Module google_api_proto::google::actions::sdk::v2

Modules

account_linking

Nested message and enum types in AccountLinking.

actions

Nested message and enum types in Actions.

actions_sdk_client

Generated client implementations.

actions_testing_client

Generated client implementations.

capability_requirement

Nested message and enum types in CapabilityRequirement.

config_file

Nested message and enum types in ConfigFile.

conversation

device_properties

Nested message and enum types in DeviceProperties.

execution_event

Nested message and enum types in ExecutionEvent.

files

Nested message and enum types in Files.

interactionmodel

settings

Nested message and enum types in Settings.

theme_customization

Nested message and enum types in ThemeCustomization.

user_input

Nested message and enum types in UserInput.

validation_result

Nested message and enum types in ValidationResult.

version

Nested message and enum types in Version.

webhook

Nested message and enum types in Webhook.

write_preview_request

Nested message and enum types in WritePreviewRequest.

Structs

AccountLinking

AccountLinking allows Google to guide the user to sign-in to the App’s web services.

AccountLinkingSecret

Information about the encrypted OAuth client secret used in account linking flows (for AUTH_CODE grant type).

Actions

Represents the list of Actions defined in a project.

CapabilityRequirement

Represents a requirement about the availability of a given capability.

Condition

Evaluated condition.

ConditionsEvaluated

Results of conditions evaluation: https://developers.google.com/assistant/conversational/scenes#conditions

ConfigFile

Represents a single file which contains structured data. Developers can define most of their project using structured config including Actions, Settings, Fulfillment.

ConfigFiles

Wrapper for repeated config files. Repeated fields cannot exist in a oneof.

CreateVersionRequest

Streaming RPC request for CreateVersion.

DataFile

Represents a single file which contains unstructured data. Examples include image files, audio files, and cloud function source code.

DataFiles

Wrapper for repeated data file. Repeated fields cannot exist in a oneof.

DecryptSecretRequest

RPC request for DecryptSecret.

DecryptSecretResponse

RPC response for DecryptSecret.

DeviceProperties

Properties of device relevant to a conversation round.

Diagnostics

Diagnostics information related to the conversation round.

Draft

Definition of draft resource.

EncryptSecretRequest

RPC request for EncryptSecret.

EncryptSecretResponse

RPC response for EncryptSecret.

EndConversation

Event which informs that conversation with agent was ended.

ExecutionEvent

Contains information about execution event which happened during processing Actions Builder conversation request. For an overview of the stages involved in a conversation request, see https://developers.google.com/assistant/conversational/actions.

ExecutionState

Current state of the execution.

Files

Wrapper for a list of files.

FormFilled

Event which happens when form is fully filled: https://developers.google.com/assistant/conversational/scenes#slot_filling

IntentMatch

Information about triggered intent match (global or within a scene): https://developers.google.com/assistant/conversational/intents

ListReleaseChannelsRequest

RPC request for listing release channels

ListReleaseChannelsResponse

RPC response for listing release channels

ListSampleProjectsRequest

RPC request for ListSampleProjects.

ListSampleProjectsResponse

RPC response for ListSampleProjects.

ListVersionsRequest

RPC request for listing versions

ListVersionsResponse

RPC response for listing versions

LocalizedSettings

Represents settings of an Actions project that are specific to a user locale. In this instance, user means the end user who invokes your Actions. This message is localizable.

Location

Container that represents a location.

Manifest

Contains information that’s “transportable” i.e. not specific to any given project and can be moved between projects.

MatchIntentsRequest

Request for finding matching intents.

MatchIntentsResponse

Response for finding matching intents.

OnSceneEnter

Information about execution of onSceneEnter stage: https://developers.google.com/assistant/conversational/scenes#on_enter

Output

User-visible output to the conversation round.

Preview

Definition of preview resource.

ReadDraftRequest

RPC request for ReadDraft.

ReadDraftResponse

Streaming RPC response for ReadDraft.

ReadVersionRequest

RPC request for ReadVersion.

ReadVersionResponse

Streaming RPC response for ReadVersion.

ReleaseChannel

Definition of release channel resource.

SampleProject

Definition of sample project resource.

SendInteractionRequest

Request for playing a round of the conversation.

SendInteractionResponse

Response to a round of the conversation.

SetWebAndAppActivityControlRequest

Request for setting Web & App Activity preferences.

Settings

Represents settings of an Actions project that are not locale specific.

SlotMatch

Information about matched slot(s): https://developers.google.com/assistant/conversational/scenes#slot_filling

SlotRequested

Information about currently requested slot: https://developers.google.com/assistant/conversational/scenes#slot_filling

SlotValidated

Event which happens after webhook validation was finished for slot(s): https://developers.google.com/assistant/conversational/scenes#slot_filling

Slots

Represents the current state of a the scene’s slots.

SurfaceRequirements

Contains a set of requirements that the client surface must support to invoke Actions in your project.

ThemeCustomization

Styles applied to cards that are presented to users

UserConversationInput

Information related to user input.

UserInput

User input provided on a conversation round.

ValidationResult

Represents a validation result associated with the app content.

ValidationResults

Wrapper for repeated validation result.

Version

Definition of version resource.

WaitingForUserInput

Event which happens when system needs user input: https://developers.google.com/assistant/conversational/scenes#input

Webhook

Metadata for different types of webhooks. If you’re using inlineCloudFunction, your source code must be in a directory with the same name as the value for the executeFunction key. For example, a value of my_webhook for theexecuteFunction key would have a code structure like this:

WebhookInitiatedTransition

Event triggered by destination scene returned from webhook: https://developers.google.com/assistant/conversational/webhooks#transition_scenes

WebhookRequest

Information about a request dispatched to the Action webhook: https://developers.google.com/assistant/conversational/webhooks#payloads

WebhookResponse

Information about a response received from the Action webhook: https://developers.google.com/assistant/conversational/webhooks#payloads

WriteDraftRequest

Streaming RPC request for WriteDraft.

WritePreviewRequest

Streaming RPC request for WritePreview.