Module google_api_proto::google::actions::sdk::v2::actions::engagement

Nested message and enum types in Engagement.

Structs

ActionLink

Indicates whether sharing links is enabled for this action and the corresponding settings. Action links are used to deep link a user into a specific action. ActionLink is deprecated. Use AssistantLink instead.

AssistantLink

Indicates whether sharing links is enabled for this action and the corresponding settings. Assistant links are used to deep link a user into a specific action.

DailyUpdate

Defines daily update settings that this engagement supports.

PushNotification

Defines push notification settings that this engagement supports.

Enums

RecurringUpdate

Recurring update settings that this engagement supports.