Module google_api_proto::google::ads::admanager::v1

Modules

ad_partner_service_client

Generated client implementations.

ad_unit

Nested message and enum types in AdUnit.

ad_unit_service_client

Generated client implementations.

applied_adsense_enabled_enum

Nested message and enum types in AppliedAdsenseEnabledEnum.

company_credit_status_enum

Nested message and enum types in CompanyCreditStatusEnum.

company_service_client

Generated client implementations.

company_type_enum

Nested message and enum types in CompanyTypeEnum.

computed_status_enum

Nested message and enum types in ComputedStatusEnum.

contact_service_client

Generated client implementations.

creative_rotation_type_enum

Nested message and enum types in CreativeRotationTypeEnum.

creative_service_client

Generated client implementations.

custom_field_data_type_enum

Nested message and enum types in CustomFieldDataTypeEnum.

custom_field_entity_type_enum

Nested message and enum types in CustomFieldEntityTypeEnum.

custom_field_service_client

Generated client implementations.

custom_field_status_enum

Nested message and enum types in CustomFieldStatusEnum.

custom_field_visibility_enum

Nested message and enum types in CustomFieldVisibilityEnum.

custom_targeting_key_reportable_type_enum

Nested message and enum types in CustomTargetingKeyReportableTypeEnum.

custom_targeting_key_service_client

Generated client implementations.

custom_targeting_key_status_enum

Nested message and enum types in CustomTargetingKeyStatusEnum.

custom_targeting_key_type_enum

Nested message and enum types in CustomTargetingKeyTypeEnum.

custom_targeting_value_match_type_enum

Nested message and enum types in CustomTargetingValueMatchTypeEnum.

custom_targeting_value_service_client

Generated client implementations.

custom_targeting_value_status_enum

Nested message and enum types in CustomTargetingValueStatusEnum.

declaration_type_enum

Nested message and enum types in DeclarationTypeEnum.

delivery_rate_type_enum

Nested message and enum types in DeliveryRateTypeEnum.

environment_type_enum

Nested message and enum types in EnvironmentTypeEnum.

export_saved_report_request

Nested message and enum types in ExportSavedReportRequest.

goal_type_enum

Nested message and enum types in GoalTypeEnum.

label_service_client

Generated client implementations.

line_item_cost_type_enum

Nested message and enum types in LineItemCostTypeEnum.

line_item_discount_type_enum

Nested message and enum types in LineItemDiscountTypeEnum.

line_item_service_client

Generated client implementations.

line_item_type_enum

Nested message and enum types in LineItemTypeEnum.

network_service_client

Generated client implementations.

order

Nested message and enum types in Order.

order_service_client

Generated client implementations.

placement_service_client

Generated client implementations.

placement_status_enum

Nested message and enum types in PlacementStatusEnum.

report_service_client

Generated client implementations.

reservation_status_enum

Nested message and enum types in ReservationStatusEnum.

role_service_client

Generated client implementations.

size_type_enum

Nested message and enum types in SizeTypeEnum.

smart_size_mode_enum

Nested message and enum types in SmartSizeModeEnum.

target_window_enum

Nested message and enum types in TargetWindowEnum.

team_service_client

Generated client implementations.

time_unit_enum

Nested message and enum types in TimeUnitEnum.

unit_type_enum

Nested message and enum types in UnitTypeEnum.

user_service_client

Generated client implementations.

Structs

AdManagerError

/ AdManagerError contains all the information required for processing a / particular error thrown by the AdManager API. / / At least one AdManagerError should be included in all error messages sent to / the client.

AdPartner

The AdPartner resource.

AdPartnerDeclaration

Represents a set of declarations about what (if any) ad partners are associated with a given creative. This can be set at the network level, as a default for all creatives, or overridden for a particular creative.

AdUnit

The AdUnit resource.

AdUnitParent

The summary of a parent AdUnit.

AdUnitSize

Represents the size, environment, and companions of an ad in an ad unit.

AppliedAdsenseEnabledEnum

Wrapper message for [AppliedAdsenseEnabled][google.ads.admanager.v1.AppliedAdsenseEnabledEnum.AppliedAdsenseEnabled]

AppliedLabel

Represents a Label that can be applied to an entity.

Company

The Company resource.

CompanyCreditStatusEnum

Wrapper message for [CompanyCreditStatus][google.ads.admanager.v1.CompanyCreditStatusEnum.CompanyCreditStatus]

CompanyTypeEnum

Wrapper message for [CompanyType][google.ads.admanager.v1.CompanyTypeEnum.CompanyType]

ComputedStatusEnum

Wrapper message for [ComputedStatus][google.ads.admanager.v1.ComputedStatusEnum.ComputedStatus].

Contact

The Contact resource.

Creative

The Creative resource.

CreativePlaceholder

Describes a slot that a creative is expected to fill. This is used in forecasting and to validate that the correct creatives are associated with the line item.

CreativeRotationTypeEnum

Wrapper message for [CreativeRotationType][google.ads.admanager.v1.CreativeRotationTypeEnum.CreativeRotationType].

CustomField

The CustomField resource.

CustomFieldDataTypeEnum

Wrapper message for [CustomFieldDataType][google.ads.admanager.v1.CustomFieldDataTypeEnum.CustomFieldDataType]

CustomFieldEntityTypeEnum

Wrapper message for [CustomFieldEntityType][google.ads.admanager.v1.CustomFieldEntityTypeEnum.CustomFieldEntityType]

CustomFieldOption

An option for a drop-down CustomField.

CustomFieldStatusEnum

Wrapper message for [CustomFieldStatus][google.ads.admanager.v1.CustomFieldStatusEnum.CustomFieldStatus]

CustomFieldVisibilityEnum

Wrapper message for [CustomFieldVisibility][google.ads.admanager.v1.CustomFieldVisibilityEnum.CustomFieldVisibility]

CustomTargetingKey

The CustomTargetingKey resource.

CustomTargetingKeyReportableTypeEnum

Wrapper message for [CustomTargetingKeyReportableType][google.ads.admanager.v1.CustomTargetingKeyReportableTypeEnum.CustomTargetingKeyReportableType]

CustomTargetingKeyStatusEnum

Wrapper message for [CustomTargetingKeyStatus][google.ads.admanager.v1.CustomTargetingKeyStatusEnum.CustomTargetingKeyStatus]

CustomTargetingKeyTypeEnum

Wrapper message for [CustomTargetingKeyType][google.ads.admanager.v1.CustomTargetingKeyTypeEnum.CustomTargetingKeyType]

CustomTargetingValue

The CustomTargetingValue resource.

CustomTargetingValueMatchTypeEnum

Wrapper message for [CustomTargetingValueMatchType][google.ads.admanager.v1.CustomTargetingValueMatchTypeEnum.CustomTargetingValueMatchType]

CustomTargetingValueStatusEnum

Wrapper message for [CustomTargetingValueStatus][google.ads.admanager.v1.CustomTargetingValueStatusEnum.CustomTargetingValueStatus]

DeclarationTypeEnum

Wrapper message for [DeclarationTypeEnum][google.ads.admanager.v1.DeclarationTypeEnum].

DeliveryRateTypeEnum

Wrapper message for [DeliveryRateType][google.ads.admanager.v1.DeliveryRateTypeEnum.DeliveryRateType].

EnvironmentTypeEnum

Wrapper message for [EnvironmentType][google.ads.admanager.v1.EnvironmentTypeEnum.EnvironmentType].

ExportSavedReportMetadata

The message stored in the google.longrunning.Operation.metadata field. Contains metadata regarding this execution.

ExportSavedReportRequest

Request proto for the configuration of a report run.

ExportSavedReportResponse

Message included in the longrunning Operation result.response field when the report completes successfully.

FrequencyCap

Represents a Frequency Cap that can be applied to an entity.

GetAdPartnerRequest

Request object for GetAdPartner method.

GetAdUnitRequest

Request object for GetAdUnit method.

GetCompanyRequest

Request object for GetCompany method.

GetContactRequest

Request object for GetContact method.

GetCreativeRequest

Request object for GetCreative method.

GetCustomFieldRequest

Request object for GetCustomField method.

GetCustomTargetingKeyRequest

Request object for GetCustomTargetingKey method.

GetCustomTargetingValueRequest

Request object for GetCustomTargetingValue method.

GetLabelRequest

Request object for GetLabel method.

GetLineItemRequest

Request object for GetLineItem method.

GetNetworkRequest

Request to get Network

GetOrderRequest

Request object for GetOrder method.

GetPlacementRequest

Request object for GetPlacement method.

GetRoleRequest

Request object for GetRole method.

GetTeamRequest

Request object for GetTeam method.

GetUserRequest

Request object for GetUser method.

Goal

Defines the criteria a [LineItem][google.ads.admanager.v1.LineItem] needs to satisfy to meet its delivery goal.

GoalTypeEnum

Wrapper message for [GoalType][google.ads.admanager.v1.GoalTypeEnum.GoalType].

Label

The Label resource.

LabelFrequencyCap

Frequency cap using a label.

LineItem

The LineItem resource.

LineItemCostTypeEnum

Wrapper message for [LineItemCostType][google.ads.admanager.v1.LineItemCostTypeEnum.LineItemCostType].

LineItemDiscountTypeEnum

Wrapper message for [LineItemDiscountType][google.ads.admanager.v1.LineItemDiscountTypeEnum.LineItemDiscountType].

LineItemTypeEnum

Wrapper message for [LineItemType][google.ads.admanager.v1.LineItemTypeEnum.LineItemType].

ListAdPartnersRequest

Request object for ListAdPartners method.

ListAdPartnersResponse

Response object for ListAdPartnersRequest containing matching AdPartner resources.

ListAdUnitsRequest

Request object for ListAdUnits method.

ListAdUnitsResponse

Response object for ListAdUnitsRequest containing matching AdUnit resources.

ListCompaniesRequest

Request object for ListCompanies method.

ListCompaniesResponse

Response object for ListCompaniesRequest containing matching Company resources.

ListContactsRequest

Request object for ListContacts method.

ListContactsResponse

Response object for ListContactsRequest containing matching Contact resources.

ListCreativesRequest

Request object for ListCreatives method.

ListCreativesResponse

Response object for ListCreativesRequest containing matching Creative resources.

ListCustomFieldsRequest

Request object for ListCustomFields method.

ListCustomFieldsResponse

Response object for ListCustomFieldsRequest containing matching CustomField objects.

ListCustomTargetingKeysRequest

Request object for ListCustomTargetingKeys method.

ListCustomTargetingKeysResponse

Response object for ListCustomTargetingKeysRequest containing matching CustomTargetingKey objects.

ListCustomTargetingValuesRequest

Request object for ListCustomTargetingValues method.

ListCustomTargetingValuesResponse

Response object for ListCustomTargetingValuesRequest containing matching CustomTargetingValue objects.

ListLabelsRequest

Request object for ListLabels method.

ListLabelsResponse

Response object for ListLabelsRequest containing matching Label resources.

ListLineItemsRequest

Request object for ListLineItems method.

ListLineItemsResponse

Response object for ListLineItemsRequest containing matching LineItem resources.

ListOrdersRequest

Request object for ListOrders method.

ListOrdersResponse

Response object for ListOrdersRequest containing matching Order resources.

ListPlacementsRequest

Request object for ListPlacements method.

ListPlacementsResponse

Response object for ListPlacementsRequest containing matching Placement objects.

ListRolesRequest

Request object for ListRoles method.

ListRolesResponse

Response object for ListRolesRequest containing matching Role resources.

ListTeamsRequest

Request object for ListTeams method.

ListTeamsResponse

Response object for ListTeamsRequest containing matching Team resources.

ListUsersRequest

Request object for ListUsers method.

ListUsersResponse

Response object for ListUsersRequest containing matching User resources.

Network

The Network resource.

Order

The Order resource.

Placement

The Placement resource.

PlacementStatusEnum

Wrapper message for [PlacementStatus][google.ads.admanager.v1.PlacementStatusEnum.PlacementStatus]

Report

The Report resource.

ReservationStatusEnum

Wrapper message for [ReservationStatus][google.ads.admanager.v1.ReservationStatusEnum.ReservationStatus].

Role

The Role resource.

Size

Represents the dimensions of an AdUnit, LineItem, or Creative.

SizeTypeEnum

Wrapper message for [SizeType][google.ads.admanager.v1.SizeTypeEnum.SizeType].

SmartSizeModeEnum

Wrapper message for [SmartSizeMode][google.ads.admanager.v1.SmartSizeModeEnum.SmartSizeMode].

TargetWindowEnum

Wrapper message for [TargetWindow][google.ads.admanager.v1.TargetWindowEnum.TargetWindow].

Team

The Team resource.

TimeUnitEnum

Wrapper message for TimeUnit.

UnitTypeEnum

Wrapper message for [UnitType][google.ads.admanager.v1.UnitTypeEnum.UnitType].

User

The User resource.