Enum google_api_proto::google::ads::googleads::v10::errors::feed_item_validation_error_enum::FeedItemValidationError

#[repr(i32)]
pub enum FeedItemValidationError {

    Unspecified,
    Unknown,
    StringTooShort,
    StringTooLong,
    ValueNotSpecified,
    InvalidDomesticPhoneNumberFormat,
    InvalidPhoneNumber,
    PhoneNumberNotSupportedForCountry,
    PremiumRateNumberNotAllowed,
    DisallowedNumberType,
    ValueOutOfRange,
    CalltrackingNotSupportedForCountry,
    CustomerNotInAllowlistForCalltracking,
    InvalidCountryCode,
    InvalidAppId,
    MissingAttributesForFields,
    InvalidTypeId,
    InvalidEmailAddress,
    InvalidHttpsUrl,
    MissingDeliveryAddress,
    StartDateAfterEndDate,
    MissingFeedItemStartTime,
    MissingFeedItemEndTime,
    MissingFeedItemId,
    VanityPhoneNumberNotAllowed,
    InvalidReviewExtensionSnippet,
    InvalidNumberFormat,
    InvalidDateFormat,
    InvalidPriceFormat,
    UnknownPlaceholderField,
    MissingEnhancedSitelinkDescriptionLine,
    ReviewExtensionSourceIneligible,
    HyphensInReviewExtensionSnippet,
    DoubleQuotesInReviewExtensionSnippet,
    QuotesInReviewExtensionSnippet,
    InvalidFormEncodedParams,
    InvalidUrlParameterName,
    NoGeocodingResult,
    SourceNameInReviewExtensionText,
    CarrierSpecificShortNumberNotAllowed,
    InvalidPlaceholderFieldId,
    InvalidUrlTag,
    ListTooLong,
    InvalidAttributesCombination,
    DuplicateValues,
    InvalidCallConversionActionId,
    CannotSetWithoutFinalUrls,
    AppIdDoesntExistInAppStore,
    InvalidFinalUrl,
    InvalidTrackingUrl,
    InvalidFinalUrlForAppDownloadUrl,
    ListTooShort,
    InvalidUserAction,
    InvalidTypeName,
    InvalidEventChangeStatus,
    InvalidSnippetsHeader,
    InvalidAndroidAppLink,
    NumberTypeWithCalltrackingNotSupportedForCountry,
    ReservedKeywordOther,
    DuplicateOptionLabels,
    DuplicateOptionPrefills,
    UnequalListLengths,
    InconsistentCurrencyCodes,
    PriceExtensionHasDuplicatedHeaders,
    ItemHasDuplicatedHeaderAndDescription,
    PriceExtensionHasTooFewItems,
    UnsupportedValue,
    InvalidFinalMobileUrl,
    InvalidKeywordlessAdRuleLabel,
    ValueTrackParameterNotSupported,
    UnsupportedValueInSelectedLanguage,
    InvalidIosAppLink,
    MissingIosAppLinkOrIosAppStoreId,
    PromotionInvalidTime,
    PromotionCannotSetPercentOffAndMoneyAmountOff,
    PromotionCannotSetPromotionCodeAndOrdersOverAmount,
    TooManyDecimalPlacesSpecified,
    AdCustomizersNotAllowed,
    InvalidLanguageCode,
    UnsupportedLanguage,
    IfFunctionNotAllowed,
    InvalidFinalUrlSuffix,
    InvalidTagInFinalUrlSuffix,
    InvalidFinalUrlSuffixFormat,
    CustomerConsentForCallRecordingRequired,
    OnlyOneDeliveryOptionIsAllowed,
    NoDeliveryOptionIsSet,
    InvalidConversionReportingState,
    ImageSizeWrong,
    EmailDeliveryNotAvailableInCountry,
    AutoReplyNotAvailableInCountry,
    InvalidLatitudeValue,
    InvalidLongitudeValue,
    TooManyLabels,
    InvalidImageUrl,
    MissingLatitudeValue,
    MissingLongitudeValue,
    AddressNotFound,
    AddressNotTargetable,
    InvalidAssetId,
    IncompatibleAssetType,
    ImageErrorUnexpectedSize,
    ImageErrorAspectRatioNotAllowed,
    ImageErrorFileTooLarge,
    ImageErrorFormatNotAllowed,
    ImageErrorConstraintsViolated,
    ImageErrorServerError,

}

The possible validation errors of a feed item.

Variants

Unspecified

No value has been specified.

Unknown

Used for return value only. Represents value unknown in this version.

StringTooShort

String is too short.

StringTooLong

String is too long.

ValueNotSpecified

Value is not provided.

InvalidDomesticPhoneNumberFormat

Phone number format is invalid for region.

InvalidPhoneNumber

String does not represent a phone number.

PhoneNumberNotSupportedForCountry

Phone number format is not compatible with country code.

PremiumRateNumberNotAllowed

Premium rate number is not allowed.

DisallowedNumberType

Phone number type is not allowed.

ValueOutOfRange

Specified value is outside of the valid range.

CalltrackingNotSupportedForCountry

Call tracking is not supported in the selected country.

CustomerNotInAllowlistForCalltracking

Customer is not on the allow-list for call tracking.

InvalidCountryCode

Country code is invalid.

InvalidAppId

The specified mobile app id is invalid.

MissingAttributesForFields

Some required field attributes are missing.

InvalidTypeId

Invalid email button type for email extension.

InvalidEmailAddress

Email address is invalid.

InvalidHttpsUrl

The HTTPS URL in email extension is invalid.

MissingDeliveryAddress

Delivery address is missing from email extension.

StartDateAfterEndDate

FeedItem scheduling start date comes after end date.

MissingFeedItemStartTime

FeedItem scheduling start time is missing.

MissingFeedItemEndTime

FeedItem scheduling end time is missing.

MissingFeedItemId

Cannot compute system attributes on a FeedItem that has no FeedItemId.

VanityPhoneNumberNotAllowed

Call extension vanity phone numbers are not supported.

InvalidReviewExtensionSnippet

Invalid review text.

InvalidNumberFormat

Invalid format for numeric value in ad parameter.

InvalidDateFormat

Invalid format for date value in ad parameter.

InvalidPriceFormat

Invalid format for price value in ad parameter.

UnknownPlaceholderField

Unrecognized type given for value in ad parameter.

MissingEnhancedSitelinkDescriptionLine

Enhanced sitelinks must have both description lines specified.

ReviewExtensionSourceIneligible

Review source is ineligible.

HyphensInReviewExtensionSnippet

Review text cannot contain hyphens or dashes.

DoubleQuotesInReviewExtensionSnippet

Review text cannot contain double quote characters.

QuotesInReviewExtensionSnippet

Review text cannot contain quote characters.

InvalidFormEncodedParams

Parameters are encoded in the wrong format.

InvalidUrlParameterName

URL parameter name must contain only letters, numbers, underscores, and dashes.

NoGeocodingResult

Cannot find address location.

SourceNameInReviewExtensionText

Review extension text has source name.

CarrierSpecificShortNumberNotAllowed

Some phone numbers can be shorter than usual. Some of these short numbers are carrier-specific, and we disallow those in ad extensions because they will not be available to all users.

InvalidPlaceholderFieldId

Triggered when a request references a placeholder field id that does not exist.

InvalidUrlTag

URL contains invalid ValueTrack tags or format.

ListTooLong

Provided list exceeds acceptable size.

InvalidAttributesCombination

Certain combinations of attributes aren’t allowed to be specified in the same feed item.

DuplicateValues

An attribute has the same value repeatedly.

InvalidCallConversionActionId

Advertisers can link a conversion action with a phone number to indicate that sufficiently long calls forwarded to that phone number should be counted as conversions of the specified type. This is an error message indicating that the conversion action specified is invalid (e.g., the conversion action does not exist within the appropriate Google Ads account, or it is a type of conversion not appropriate to phone call conversions).

CannotSetWithoutFinalUrls

Tracking template requires final url to be set.

AppIdDoesntExistInAppStore

An app id was provided that doesn’t exist in the given app store.

InvalidFinalUrl

Invalid U2 final url.

InvalidTrackingUrl

Invalid U2 tracking url.

InvalidFinalUrlForAppDownloadUrl

Final URL should start from App download URL.

ListTooShort

List provided is too short.

InvalidUserAction

User Action field has invalid value.

InvalidTypeName

Type field has invalid value.

InvalidEventChangeStatus

Change status for event is invalid.

InvalidSnippetsHeader

The header of a structured snippets extension is not one of the valid headers.

InvalidAndroidAppLink

Android app link is not formatted correctly

NumberTypeWithCalltrackingNotSupportedForCountry

Phone number incompatible with call tracking for country.

ReservedKeywordOther

The input is identical to a reserved keyword

DuplicateOptionLabels

Each option label in the message extension must be unique.

DuplicateOptionPrefills

Each option prefill in the message extension must be unique.

UnequalListLengths

In message extensions, the number of optional labels and optional prefills must be the same.

InconsistentCurrencyCodes

All currency codes in an ad extension must be the same.

PriceExtensionHasDuplicatedHeaders

Headers in price extension are not unique.

ItemHasDuplicatedHeaderAndDescription

Header and description in an item are the same.

PriceExtensionHasTooFewItems

Price extension has too few items.

UnsupportedValue

The given value is not supported.

InvalidFinalMobileUrl

Invalid final mobile url.

InvalidKeywordlessAdRuleLabel

The given string value of Label contains invalid characters

ValueTrackParameterNotSupported

The given URL contains value track parameters.

UnsupportedValueInSelectedLanguage

The given value is not supported in the selected language of an extension.

InvalidIosAppLink

The iOS app link is not formatted correctly.

MissingIosAppLinkOrIosAppStoreId

iOS app link or iOS app store id is missing.

PromotionInvalidTime

Promotion time is invalid.

PromotionCannotSetPercentOffAndMoneyAmountOff

Both the percent off and money amount off fields are set.

PromotionCannotSetPromotionCodeAndOrdersOverAmount

Both the promotion code and orders over amount fields are set.

TooManyDecimalPlacesSpecified

Too many decimal places are specified.

AdCustomizersNotAllowed

Ad Customizers are present and not allowed.

InvalidLanguageCode

Language code is not valid.

UnsupportedLanguage

Language is not supported.

IfFunctionNotAllowed

IF Function is present and not allowed.

InvalidFinalUrlSuffix

Final url suffix is not valid.

InvalidTagInFinalUrlSuffix

Final url suffix contains an invalid tag.

InvalidFinalUrlSuffixFormat

Final url suffix is formatted incorrectly.

CustomerConsentForCallRecordingRequired

Consent for call recording, which is required for the use of call extensions, was not provided by the advertiser. Please see https://support.google.com/google-ads/answer/7412639.

OnlyOneDeliveryOptionIsAllowed

Multiple message delivery options are set.

NoDeliveryOptionIsSet

No message delivery option is set.

InvalidConversionReportingState

String value of conversion reporting state field is not valid.

ImageSizeWrong

Image size is not right.

EmailDeliveryNotAvailableInCountry

Email delivery is not supported in the country specified in the country code field.

AutoReplyNotAvailableInCountry

Auto reply is not supported in the country specified in the country code field.

InvalidLatitudeValue

Invalid value specified for latitude.

InvalidLongitudeValue

Invalid value specified for longitude.

TooManyLabels

Too many label fields provided.

InvalidImageUrl

Invalid image url.

MissingLatitudeValue

Latitude value is missing.

MissingLongitudeValue

Longitude value is missing.

AddressNotFound

Unable to find address.

AddressNotTargetable

Cannot target provided address.

InvalidAssetId

The specified asset ID does not exist.

IncompatibleAssetType

The asset type cannot be set for the field.

ImageErrorUnexpectedSize

The image has unexpected size.

ImageErrorAspectRatioNotAllowed

The image aspect ratio is not allowed.

ImageErrorFileTooLarge

The image file is too large.

ImageErrorFormatNotAllowed

The image format is unsupported.

ImageErrorConstraintsViolated

Image violates constraints without more details.

ImageErrorServerError

An error occurred when validating image.

Implementations

impl FeedItemValidationError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of FeedItemValidationError.

pub fn from_i32(value: i32) -> Option<FeedItemValidationError>

Converts an i32 to a FeedItemValidationError, or None if value is not a valid variant.