Struct google_api_proto::google::ads::googleads::v12::common::PromotionFeedItem

pub struct PromotionFeedItem {

    pub promotion_target: Option<String>,
    pub discount_modifier: i32,
    pub promotion_start_date: Option<String>,
    pub promotion_end_date: Option<String>,
    pub occasion: i32,
    pub final_urls: Vec<String>,
    pub final_mobile_urls: Vec<String>,
    pub tracking_url_template: Option<String>,
    pub url_custom_parameters: Vec<CustomParameter>,
    pub final_url_suffix: Option<String>,
    pub language_code: Option<String>,
    pub discount_type: Option<DiscountType>,
    pub promotion_trigger: Option<PromotionTrigger>,

}

Represents a Promotion extension.

Fields
promotion_target: Option<String>

A freeform description of what the promotion is targeting. This field is required.

discount_modifier: i32

Enum that modifies the qualification of the discount.

promotion_start_date: Option<String>

Start date of when the promotion is eligible to be redeemed.

promotion_end_date: Option<String>

Last date when the promotion is eligible to be redeemed.

occasion: i32

The occasion the promotion was intended for. If an occasion is set, the redemption window will need to fall within the date range associated with the occasion.

final_urls: Vec<String>

A list of possible final URLs after all cross domain redirects. This field is required.

final_mobile_urls: Vec<String>

A list of possible final mobile URLs after all cross domain redirects.

tracking_url_template: Option<String>

URL template for constructing a tracking URL.

url_custom_parameters: Vec<CustomParameter>

A list of mappings to be used for substituting URL custom parameter tags in the tracking_url_template, final_urls, and/or final_mobile_urls.

final_url_suffix: Option<String>

URL template for appending params to landing page URLs served with parallel tracking.

language_code: Option<String>

The language of the promotion. Represented as BCP 47 language tag.

discount_type: Option<DiscountType>

Discount type, can be percentage off or amount off.

promotion_trigger: Option<PromotionTrigger>

Promotion trigger. Can be by promotion code or promo by eligible order amount.

Implementations

impl PromotionFeedItem

pub fn discount_modifier(&self) -> PromotionExtensionDiscountModifier

Returns the enum value of discount_modifier, or the default if the field is set to an invalid enum value.

pub fn set_discount_modifier( &mut self, value: PromotionExtensionDiscountModifier )

Sets discount_modifier to the provided enum value.

pub fn occasion(&self) -> PromotionExtensionOccasion

Returns the enum value of occasion, or the default if the field is set to an invalid enum value.

pub fn set_occasion(&mut self, value: PromotionExtensionOccasion)

Sets occasion to the provided enum value.

pub fn promotion_target(&self) -> &str

Returns the value of promotion_target, or the default value if promotion_target is unset.

pub fn promotion_start_date(&self) -> &str

Returns the value of promotion_start_date, or the default value if promotion_start_date is unset.

pub fn promotion_end_date(&self) -> &str

Returns the value of promotion_end_date, or the default value if promotion_end_date is unset.

pub fn tracking_url_template(&self) -> &str

Returns the value of tracking_url_template, or the default value if tracking_url_template is unset.

pub fn final_url_suffix(&self) -> &str

Returns the value of final_url_suffix, or the default value if final_url_suffix is unset.

pub fn language_code(&self) -> &str

Returns the value of language_code, or the default value if language_code is unset.