Enum google_api_proto::google::ads::googleads::v12::errors::user_list_error_enum::UserListError

#[repr(i32)]
pub enum UserListError {

    Unspecified,
    Unknown,
    ExternalRemarketingUserListMutateNotSupported,
    ConcreteTypeRequired,
    ConversionTypeIdRequired,
    DuplicateConversionTypes,
    InvalidConversionType,
    InvalidDescription,
    InvalidName,
    InvalidType,
    CanNotAddLogicalListAsLogicalListOperand,
    InvalidUserListLogicalRuleOperand,
    NameAlreadyUsed,
    NewConversionTypeNameRequired,
    ConversionTypeNameAlreadyUsed,
    OwnershipRequiredForSet,
    UserListMutateNotSupported,
    InvalidRule,
    InvalidDateRange,
    CanNotMutateSensitiveUserlist,
    MaxNumRulebasedUserlists,
    CannotModifyBillableRecordCount,
    AppIdNotSet,
    UserlistNameIsReservedForSystemList,
    AdvertiserNotOnAllowlistForUsingUploadedData,
    RuleTypeIsNotSupported,
    CanNotAddASimilarUserlistAsLogicalListOperand,
    CanNotMixCrmBasedInLogicalListWithOtherLists,

}

Enum describing possible user list errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

ExternalRemarketingUserListMutateNotSupported

Creating and updating external remarketing user lists is not supported.

ConcreteTypeRequired

Concrete type of user list is required.

ConversionTypeIdRequired

Creating/updating user list conversion types requires specifying the conversion type Id.

DuplicateConversionTypes

Remarketing user list cannot have duplicate conversion types.

InvalidConversionType

Conversion type is invalid/unknown.

InvalidDescription

User list description is empty or invalid.

InvalidName

User list name is empty or invalid.

InvalidType

Type of the UserList does not match.

CanNotAddLogicalListAsLogicalListOperand

Embedded logical user lists are not allowed.

InvalidUserListLogicalRuleOperand

User list rule operand is invalid.

NameAlreadyUsed

Name is already being used for another user list for the account.

NewConversionTypeNameRequired

Name is required when creating a new conversion type.

ConversionTypeNameAlreadyUsed

The given conversion type name has been used.

OwnershipRequiredForSet

Only an owner account may edit a user list.

UserListMutateNotSupported

Creating user list without setting type in oneof user_list field, or creating/updating read-only user list types is not allowed.

InvalidRule

Rule is invalid.

InvalidDateRange

The specified date range is empty.

CanNotMutateSensitiveUserlist

A UserList which is privacy sensitive or legal rejected cannot be mutated by external users.

MaxNumRulebasedUserlists

Maximum number of rulebased user lists a customer can have.

CannotModifyBillableRecordCount

BasicUserList’s billable record field cannot be modified once it is set.

AppIdNotSet

crm_based_user_list.app_id field must be set when upload_key_type is MOBILE_ADVERTISING_ID.

UserlistNameIsReservedForSystemList

Name of the user list is reserved for system generated lists and cannot be used.

AdvertiserNotOnAllowlistForUsingUploadedData

Advertiser needs to be on the allow-list to use remarketing lists created from advertiser uploaded data (for example, Customer Match lists).

RuleTypeIsNotSupported

The provided rule_type is not supported for the user list.

CanNotAddASimilarUserlistAsLogicalListOperand

Similar user list cannot be used as a logical user list operand.

CanNotMixCrmBasedInLogicalListWithOtherLists

Logical user list should not have a mix of CRM based user list and other types of lists in its rules.

Implementations

impl UserListError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of UserListError.

pub fn from_i32(value: i32) -> Option<UserListError>

Converts an i32 to a UserListError, or None if value is not a valid variant.

impl UserListError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.