Struct google_api_proto::google::ads::googleads::v13::common::LegacyResponsiveDisplayAdInfo

pub struct LegacyResponsiveDisplayAdInfo {

    pub short_headline: Option<String>,
    pub long_headline: Option<String>,
    pub description: Option<String>,
    pub business_name: Option<String>,
    pub allow_flexible_color: Option<bool>,
    pub accent_color: Option<String>,
    pub main_color: Option<String>,
    pub call_to_action_text: Option<String>,
    pub logo_image: Option<String>,
    pub square_logo_image: Option<String>,
    pub marketing_image: Option<String>,
    pub square_marketing_image: Option<String>,
    pub format_setting: i32,
    pub price_prefix: Option<String>,
    pub promo_text: Option<String>,

}

A legacy responsive display ad. Ads of this type are labeled ‘Responsive ads’ in the Google Ads UI.

Fields
short_headline: Option<String>

The short version of the ad’s headline.

long_headline: Option<String>

The long version of the ad’s headline.

description: Option<String>

The description of the ad.

business_name: Option<String>

The business name in the ad.

allow_flexible_color: Option<bool>

Advertiser’s consent to allow flexible color. When true, the ad may be served with different color if necessary. When false, the ad will be served with the specified colors or a neutral color. The default value is true. Must be true if main_color and accent_color are not set.

accent_color: Option<String>

The accent color of the ad in hexadecimal, for example, #ffffff for white. If one of main_color and accent_color is set, the other is required as well.

main_color: Option<String>

The main color of the ad in hexadecimal, for example, #ffffff for white. If one of main_color and accent_color is set, the other is required as well.

call_to_action_text: Option<String>

The call-to-action text for the ad.

logo_image: Option<String>

The MediaFile resource name of the logo image used in the ad.

square_logo_image: Option<String>

The MediaFile resource name of the square logo image used in the ad.

marketing_image: Option<String>

The MediaFile resource name of the marketing image used in the ad.

square_marketing_image: Option<String>

The MediaFile resource name of the square marketing image used in the ad.

format_setting: i32

Specifies which format the ad will be served in. Default is ALL_FORMATS.

price_prefix: Option<String>

Prefix before price. For example, ‘as low as’.

promo_text: Option<String>

Promotion text used for dynamic formats of responsive ads. For example ‘Free two-day shipping’.

Implementations

impl LegacyResponsiveDisplayAdInfo

pub fn format_setting(&self) -> DisplayAdFormatSetting

Returns the enum value of format_setting, or the default if the field is set to an invalid enum value.

pub fn set_format_setting(&mut self, value: DisplayAdFormatSetting)

Sets format_setting to the provided enum value.

pub fn short_headline(&self) -> &str

Returns the value of short_headline, or the default value if short_headline is unset.

pub fn long_headline(&self) -> &str

Returns the value of long_headline, or the default value if long_headline is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn business_name(&self) -> &str

Returns the value of business_name, or the default value if business_name is unset.

pub fn allow_flexible_color(&self) -> bool

Returns the value of allow_flexible_color, or the default value if allow_flexible_color is unset.

pub fn accent_color(&self) -> &str

Returns the value of accent_color, or the default value if accent_color is unset.

pub fn main_color(&self) -> &str

Returns the value of main_color, or the default value if main_color is unset.

pub fn call_to_action_text(&self) -> &str

Returns the value of call_to_action_text, or the default value if call_to_action_text is unset.

pub fn logo_image(&self) -> &str

Returns the value of logo_image, or the default value if logo_image is unset.

pub fn square_logo_image(&self) -> &str

Returns the value of square_logo_image, or the default value if square_logo_image is unset.

pub fn marketing_image(&self) -> &str

Returns the value of marketing_image, or the default value if marketing_image is unset.

pub fn square_marketing_image(&self) -> &str

Returns the value of square_marketing_image, or the default value if square_marketing_image is unset.

pub fn price_prefix(&self) -> &str

Returns the value of price_prefix, or the default value if price_prefix is unset.

pub fn promo_text(&self) -> &str

Returns the value of promo_text, or the default value if promo_text is unset.