Enum google_api_proto::google::ads::googleads::v13::enums::bidding_strategy_type_enum::BiddingStrategyType

#[repr(i32)]
pub enum BiddingStrategyType {

    Unspecified,
    Unknown,
    Commission,
    EnhancedCpc,
    Invalid,
    ManualCpa,
    ManualCpc,
    ManualCpm,
    ManualCpv,
    MaximizeConversions,
    MaximizeConversionValue,
    PageOnePromoted,
    PercentCpc,
    TargetCpa,
    TargetCpm,
    TargetImpressionShare,
    TargetOutrankShare,
    TargetRoas,
    TargetSpend,

}

Enum describing possible bidding strategy types.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

Commission

Commission is an automatic bidding strategy in which the advertiser pays a certain portion of the conversion value.

EnhancedCpc

Enhanced CPC is a bidding strategy that raises bids for clicks that seem more likely to lead to a conversion and lowers them for clicks where they seem less likely.

Invalid

Used for return value only. Indicates that a campaign does not have a bidding strategy. This prevents the campaign from serving. For example, a campaign may be attached to a manager bidding strategy and the serving account is subsequently unlinked from the manager account. In this case the campaign will automatically be detached from the now inaccessible manager bidding strategy and transition to the INVALID bidding strategy type.

ManualCpa

Manual bidding strategy that allows advertiser to set the bid per advertiser-specified action.

ManualCpc

Manual click based bidding where user pays per click.

ManualCpm

Manual impression based bidding where user pays per thousand impressions.

ManualCpv

A bidding strategy that pays a configurable amount per video view.

MaximizeConversions

A bidding strategy that automatically maximizes number of conversions given a daily budget.

MaximizeConversionValue

An automated bidding strategy that automatically sets bids to maximize revenue while spending your budget.

PageOnePromoted

Page-One Promoted bidding scheme, which sets max cpc bids to target impressions on page one or page one promoted slots on google.com. This enum value is deprecated.

PercentCpc

Percent Cpc is bidding strategy where bids are a fraction of the advertised price for some good or service.

TargetCpa

Target CPA is an automated bid strategy that sets bids to help get as many conversions as possible at the target cost-per-acquisition (CPA) you set.

TargetCpm

Target CPM is an automated bid strategy that sets bids to help get as many impressions as possible at the target cost per one thousand impressions (CPM) you set.

TargetImpressionShare

An automated bidding strategy that sets bids so that a certain percentage of search ads are shown at the top of the first page (or other targeted location).

TargetOutrankShare

Target Outrank Share is an automated bidding strategy that sets bids based on the target fraction of auctions where the advertiser should outrank a specific competitor. This enum value is deprecated.

TargetRoas

Target ROAS is an automated bidding strategy that helps you maximize revenue while averaging a specific target Return On Average Spend (ROAS).

TargetSpend

Target Spend is an automated bid strategy that sets your bids to help get as many clicks as possible within your budget.

Implementations

impl BiddingStrategyType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of BiddingStrategyType.

pub fn from_i32(value: i32) -> Option<BiddingStrategyType>

Converts an i32 to a BiddingStrategyType, or None if value is not a valid variant.

impl BiddingStrategyType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.