Enum google_api_proto::google::ads::googleads::v13::enums::conversion_action_type_enum::ConversionActionType

#[repr(i32)]
pub enum ConversionActionType {

    Unspecified,
    Unknown,
    AdCall,
    ClickToCall,
    GooglePlayDownload,
    GooglePlayInAppPurchase,
    UploadCalls,
    UploadClicks,
    Webpage,
    WebsiteCall,
    StoreSalesDirectUpload,
    StoreSales,
    FirebaseAndroidFirstOpen,
    FirebaseAndroidInAppPurchase,
    FirebaseAndroidCustom,
    FirebaseIosFirstOpen,
    FirebaseIosInAppPurchase,
    FirebaseIosCustom,
    ThirdPartyAppAnalyticsAndroidFirstOpen,
    ThirdPartyAppAnalyticsAndroidInAppPurchase,
    ThirdPartyAppAnalyticsAndroidCustom,
    ThirdPartyAppAnalyticsIosFirstOpen,
    ThirdPartyAppAnalyticsIosInAppPurchase,
    ThirdPartyAppAnalyticsIosCustom,
    AndroidAppPreRegistration,
    AndroidInstallsAllOtherApps,
    FloodlightAction,
    FloodlightTransaction,
    GoogleHosted,
    LeadFormSubmit,
    Salesforce,
    SearchAds360,
    SmartCampaignAdClicksToCall,
    SmartCampaignMapClicksToCall,
    SmartCampaignMapDirections,
    SmartCampaignTrackedCalls,
    StoreVisits,
    WebpageCodeless,

}

Possible types of a conversion action.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

AdCall

Conversions that occur when a user clicks on an ad’s call extension.

ClickToCall

Conversions that occur when a user on a mobile device clicks a phone number.

GooglePlayDownload

Conversions that occur when a user downloads a mobile app from the Google Play Store.

GooglePlayInAppPurchase

Conversions that occur when a user makes a purchase in an app through Android billing.

UploadCalls

Call conversions that are tracked by the advertiser and uploaded.

UploadClicks

Conversions that are tracked by the advertiser and uploaded with attributed clicks.

Webpage

Conversions that occur on a webpage.

WebsiteCall

Conversions that occur when a user calls a dynamically-generated phone number from an advertiser’s website.

StoreSalesDirectUpload

Store Sales conversion based on first-party or third-party merchant data uploads. Only customers on the allowlist can use store sales direct upload types.

StoreSales

Store Sales conversion based on first-party or third-party merchant data uploads and/or from in-store purchases using cards from payment networks. Only customers on the allowlist can use store sales types. Read only.

FirebaseAndroidFirstOpen

Android app first open conversions tracked through Firebase.

FirebaseAndroidInAppPurchase

Android app in app purchase conversions tracked through Firebase.

FirebaseAndroidCustom

Android app custom conversions tracked through Firebase.

FirebaseIosFirstOpen

iOS app first open conversions tracked through Firebase.

FirebaseIosInAppPurchase

iOS app in app purchase conversions tracked through Firebase.

FirebaseIosCustom

iOS app custom conversions tracked through Firebase.

ThirdPartyAppAnalyticsAndroidFirstOpen

Android app first open conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsAndroidInAppPurchase

Android app in app purchase conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsAndroidCustom

Android app custom conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosFirstOpen

iOS app first open conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosInAppPurchase

iOS app in app purchase conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosCustom

iOS app custom conversions tracked through Third Party App Analytics.

AndroidAppPreRegistration

Conversions that occur when a user pre-registers a mobile app from the Google Play Store. Read only.

AndroidInstallsAllOtherApps

Conversions that track all Google Play downloads which aren’t tracked by an app-specific type. Read only.

FloodlightAction

Floodlight activity that counts the number of times that users have visited a particular webpage after seeing or clicking on one of an advertiser’s ads. Read only.

FloodlightTransaction

Floodlight activity that tracks the number of sales made or the number of items purchased. Can also capture the total value of each sale. Read only.

GoogleHosted

Conversions that track local actions from Google’s products and services after interacting with an ad. Read only.

LeadFormSubmit

Conversions reported when a user submits a lead form. Read only.

Salesforce

Conversions that come from Salesforce. Read only.

SearchAds360

Conversions imported from Search Ads 360 Floodlight data. Read only.

SmartCampaignAdClicksToCall

Call conversions that occur on Smart campaign Ads without call tracking setup, using Smart campaign custom criteria. Read only.

SmartCampaignMapClicksToCall

The user clicks on a call element within Google Maps. Smart campaign only. Read only.

SmartCampaignMapDirections

The user requests directions to a business location within Google Maps. Smart campaign only. Read only.

SmartCampaignTrackedCalls

Call conversions that occur on Smart campaign Ads with call tracking setup, using Smart campaign custom criteria. Read only.

StoreVisits

Conversions that occur when a user visits an advertiser’s retail store. Read only.

WebpageCodeless

Conversions created from website events (such as form submissions or page loads), that don’t use individually coded event snippets.

Implementations

impl ConversionActionType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionActionType.

pub fn from_i32(value: i32) -> Option<ConversionActionType>

Converts an i32 to a ConversionActionType, or None if value is not a valid variant.

impl ConversionActionType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.