Enum google_api_proto::google::ads::googleads::v13::enums::distance_bucket_enum::DistanceBucket

#[repr(i32)]
pub enum DistanceBucket {

    Unspecified,
    Unknown,
    Within700m,
    Within1km,
    Within5km,
    Within10km,
    Within15km,
    Within20km,
    Within25km,
    Within30km,
    Within35km,
    Within40km,
    Within45km,
    Within50km,
    Within55km,
    Within60km,
    Within65km,
    Beyond65km,
    Within07miles,
    Within1mile,
    Within5miles,
    Within10miles,
    Within15miles,
    Within20miles,
    Within25miles,
    Within30miles,
    Within35miles,
    Within40miles,
    Beyond40miles,

}

The distance bucket for a user’s distance from an advertiser’s location extension.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

Within700m

User was within 700m of the location.

Within1km

User was within 1KM of the location.

Within5km

User was within 5KM of the location.

Within10km

User was within 10KM of the location.

Within15km

User was within 15KM of the location.

Within20km

User was within 20KM of the location.

Within25km

User was within 25KM of the location.

Within30km

User was within 30KM of the location.

Within35km

User was within 35KM of the location.

Within40km

User was within 40KM of the location.

Within45km

User was within 45KM of the location.

Within50km

User was within 50KM of the location.

Within55km

User was within 55KM of the location.

Within60km

User was within 60KM of the location.

Within65km

User was within 65KM of the location.

Beyond65km

User was beyond 65KM of the location.

Within07miles

User was within 0.7 miles of the location.

Within1mile

User was within 1 mile of the location.

Within5miles

User was within 5 miles of the location.

Within10miles

User was within 10 miles of the location.

Within15miles

User was within 15 miles of the location.

Within20miles

User was within 20 miles of the location.

Within25miles

User was within 25 miles of the location.

Within30miles

User was within 30 miles of the location.

Within35miles

User was within 35 miles of the location.

Within40miles

User was within 40 miles of the location.

Beyond40miles

User was beyond 40 miles of the location.

Implementations

impl DistanceBucket

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of DistanceBucket.

pub fn from_i32(value: i32) -> Option<DistanceBucket>

Converts an i32 to a DistanceBucket, or None if value is not a valid variant.

impl DistanceBucket

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.