Enum google_api_proto::google::ads::googleads::v13::enums::lead_form_field_user_input_type_enum::LeadFormFieldUserInputType

#[repr(i32)]
pub enum LeadFormFieldUserInputType {

    Unspecified,
    Unknown,
    FullName,
    Email,
    PhoneNumber,
    PostalCode,
    StreetAddress,
    City,
    Region,
    Country,
    WorkEmail,
    CompanyName,
    WorkPhone,
    JobTitle,
    GovernmentIssuedIdCpfBr,
    GovernmentIssuedIdDniAr,
    GovernmentIssuedIdDniPe,
    GovernmentIssuedIdRutCl,
    GovernmentIssuedIdCcCo,
    GovernmentIssuedIdCiEc,
    GovernmentIssuedIdRfcMx,
    FirstName,
    LastName,
    VehicleModel,
    VehicleType,
    PreferredDealership,
    VehiclePurchaseTimeline,
    VehicleOwnership,
    VehiclePaymentType,
    VehicleCondition,
    CompanySize,
    AnnualSales,
    YearsInBusiness,
    JobDepartment,
    JobRole,
    Over18Age,
    Over19Age,
    Over20Age,
    Over21Age,
    Over22Age,
    Over23Age,
    Over24Age,
    Over25Age,
    Over26Age,
    Over27Age,
    Over28Age,
    Over29Age,
    Over30Age,
    Over31Age,
    Over32Age,
    Over33Age,
    Over34Age,
    Over35Age,
    Over36Age,
    Over37Age,
    Over38Age,
    Over39Age,
    Over40Age,
    Over41Age,
    Over42Age,
    Over43Age,
    Over44Age,
    Over45Age,
    Over46Age,
    Over47Age,
    Over48Age,
    Over49Age,
    Over50Age,
    Over51Age,
    Over52Age,
    Over53Age,
    Over54Age,
    Over55Age,
    Over56Age,
    Over57Age,
    Over58Age,
    Over59Age,
    Over60Age,
    Over61Age,
    Over62Age,
    Over63Age,
    Over64Age,
    Over65Age,
    EducationProgram,
    EducationCourse,
    Product,
    Service,
    Offer,
    Category,
    PreferredContactMethod,
    PreferredLocation,
    PreferredContactTime,
    PurchaseTimeline,
    YearsOfExperience,
    JobIndustry,
    LevelOfEducation,
    PropertyType,
    RealtorHelpGoal,
    PropertyCommunity,
    PriceRange,
    NumberOfBedrooms,
    FurnishedProperty,
    PetsAllowedProperty,
    NextPlannedPurchase,
    EventSignupInterest,
    PreferredShoppingPlaces,
    FavoriteBrand,
    TransportationCommercialLicenseType,
    EventBookingInterest,
    DestinationCountry,
    DestinationCity,
    DepartureCountry,
    DepartureCity,
    DepartureDate,
    ReturnDate,
    NumberOfTravelers,
    TravelBudget,
    TravelAccommodation,

}

Enum describing the input type of a lead form field.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

FullName

The user will be asked to fill in their given and family name. This field cannot be set at the same time as GIVEN_NAME or FAMILY_NAME.

Email

The user will be asked to fill in their email address.

PhoneNumber

The user will be asked to fill in their phone number.

PostalCode

The user will be asked to fill in their zip code.

StreetAddress

The user will be asked to fill in their street address.

City

The user will be asked to fill in their city.

Region

The user will be asked to fill in their region part of the address (for example, state for US, province for Canada).

Country

The user will be asked to fill in their country.

WorkEmail

The user will be asked to fill in their work email address.

CompanyName

The user will be asked to fill in their company name.

WorkPhone

The user will be asked to fill in their work phone.

JobTitle

The user will be asked to fill in their job title.

GovernmentIssuedIdCpfBr

The user will be asked to fill in their CPF for Brazil users.

GovernmentIssuedIdDniAr

The user will be asked to fill in their DNI for Argentina users.

GovernmentIssuedIdDniPe

The user will be asked to fill in their DNI for Peru users.

GovernmentIssuedIdRutCl

The user will be asked to fill in their RUT for Chile users.

GovernmentIssuedIdCcCo

The user will be asked to fill in their CC for Colombia users.

GovernmentIssuedIdCiEc

The user will be asked to fill in their CI for Ecuador users.

GovernmentIssuedIdRfcMx

The user will be asked to fill in their RFC for Mexico users.

FirstName

The user will be asked to fill in their first name. This field can not be set at the same time as FULL_NAME.

LastName

The user will be asked to fill in their last name. This field can not be set at the same time as FULL_NAME.

VehicleModel

Question: “Which model are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleType

Question: “Which type of vehicle are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredDealership

Question: “What is your preferred dealership?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehiclePurchaseTimeline

Question: “When do you plan on purchasing a vehicle?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleOwnership

Question: “Do you own a vehicle?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehiclePaymentType

Question: “What vehicle ownership option are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleCondition

Question: “What type of vehicle condition are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

CompanySize

Question: “What size is your company?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

AnnualSales

Question: “What is your annual sales volume?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

YearsInBusiness

Question: “How many years have you been in business?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobDepartment

Question: “What is your job department?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobRole

Question: “What is your job role?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over18Age

Question: “Are you over 18 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over19Age

Question: “Are you over 19 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over20Age

Question: “Are you over 20 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over21Age

Question: “Are you over 21 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over22Age

Question: “Are you over 22 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over23Age

Question: “Are you over 23 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over24Age

Question: “Are you over 24 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over25Age

Question: “Are you over 25 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over26Age

Question: “Are you over 26 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over27Age

Question: “Are you over 27 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over28Age

Question: “Are you over 28 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over29Age

Question: “Are you over 29 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over30Age

Question: “Are you over 30 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over31Age

Question: “Are you over 31 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over32Age

Question: “Are you over 32 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over33Age

Question: “Are you over 33 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over34Age

Question: “Are you over 34 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over35Age

Question: “Are you over 35 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over36Age

Question: “Are you over 36 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over37Age

Question: “Are you over 37 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over38Age

Question: “Are you over 38 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over39Age

Question: “Are you over 39 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over40Age

Question: “Are you over 40 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over41Age

Question: “Are you over 41 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over42Age

Question: “Are you over 42 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over43Age

Question: “Are you over 43 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over44Age

Question: “Are you over 44 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over45Age

Question: “Are you over 45 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over46Age

Question: “Are you over 46 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over47Age

Question: “Are you over 47 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over48Age

Question: “Are you over 48 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over49Age

Question: “Are you over 49 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over50Age

Question: “Are you over 50 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over51Age

Question: “Are you over 51 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over52Age

Question: “Are you over 52 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over53Age

Question: “Are you over 53 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over54Age

Question: “Are you over 54 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over55Age

Question: “Are you over 55 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over56Age

Question: “Are you over 56 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over57Age

Question: “Are you over 57 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over58Age

Question: “Are you over 58 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over59Age

Question: “Are you over 59 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over60Age

Question: “Are you over 60 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over61Age

Question: “Are you over 61 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over62Age

Question: “Are you over 62 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over63Age

Question: “Are you over 63 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over64Age

Question: “Are you over 64 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over65Age

Question: “Are you over 65 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EducationProgram

Question: “Which program are you interested in?” Category: “Education” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EducationCourse

Question: “Which course are you interested in?” Category: “Education” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Product

Question: “Which product are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Service

Question: “Which service are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Offer

Question: “Which offer are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Category

Question: “Which category are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredContactMethod

Question: “What is your preferred method of contact?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredLocation

Question: “What is your preferred location?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredContactTime

Question: “What is the best time to contact you?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PurchaseTimeline

Question: “When are you looking to make a purchase?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

YearsOfExperience

Question: “How many years of work experience do you have?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobIndustry

Question: “What industry do you work in?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

LevelOfEducation

Question: “What is your highest level of education?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PropertyType

Question: “What type of property are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

RealtorHelpGoal

Question: “What do you need a realtor’s help with?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PropertyCommunity

Question: “What neighborhood are you interested in?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PriceRange

Question: “What price range are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NumberOfBedrooms

Question: “How many bedrooms are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

FurnishedProperty

Question: “Are you looking for a fully furnished property?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PetsAllowedProperty

Question: “Are you looking for properties that allow pets?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NextPlannedPurchase

Question: “What is the next product you plan to purchase?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EventSignupInterest

Question: “Would you like to sign up for an event?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredShoppingPlaces

Question: “Where are you interested in shopping?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

FavoriteBrand

Question: “What is your favorite brand?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TransportationCommercialLicenseType

Question: “Which type of valid commercial license do you have?” Category: “Transportation” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EventBookingInterest

Question: “Interested in booking an event?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DestinationCountry

Question: “What is your destination country?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DestinationCity

Question: “What is your destination city?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureCountry

Question: “What is your departure country?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureCity

Question: “What is your departure city?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureDate

Question: “What is your departure date?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

ReturnDate

Question: “What is your return date?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NumberOfTravelers

Question: “How many people are you traveling with?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TravelBudget

Question: “What is your travel budget?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TravelAccommodation

Question: “Where do you want to stay during your travel?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Implementations

impl LeadFormFieldUserInputType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of LeadFormFieldUserInputType.

pub fn from_i32(value: i32) -> Option<LeadFormFieldUserInputType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a LeadFormFieldUserInputType, or None if value is not a valid variant.

impl LeadFormFieldUserInputType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.