Enum google_api_proto::google::ads::googleads::v13::errors::campaign_budget_error_enum::CampaignBudgetError

#[repr(i32)]
pub enum CampaignBudgetError {

    Unspecified,
    Unknown,
    CampaignBudgetCannotBeShared,
    CampaignBudgetRemoved,
    CampaignBudgetInUse,
    CampaignBudgetPeriodNotAvailable,
    CannotModifyFieldOfImplicitlySharedCampaignBudget,
    CannotUpdateCampaignBudgetToImplicitlyShared,
    CannotUpdateCampaignBudgetToExplicitlySharedWithoutName,
    CannotUpdateCampaignBudgetToExplicitlyShared,
    CannotUseImplicitlySharedCampaignBudgetWithMultipleCampaigns,
    DuplicateName,
    MoneyAmountInWrongCurrency,
    MoneyAmountLessThanCurrencyMinimumCpc,
    MoneyAmountTooLarge,
    NegativeMoneyAmount,
    NonMultipleOfMinimumCurrencyUnit,
    TotalBudgetAmountMustBeUnsetForBudgetPeriodDaily,
    InvalidPeriod,
    CannotUseAcceleratedDeliveryMode,

}

Enum describing possible campaign budget errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

CampaignBudgetCannotBeShared

The campaign budget cannot be shared.

CampaignBudgetRemoved

The requested campaign budget no longer exists.

CampaignBudgetInUse

The campaign budget is associated with at least one campaign, and so the campaign budget cannot be removed.

CampaignBudgetPeriodNotAvailable

Customer is not on the allow-list for this campaign budget period.

CannotModifyFieldOfImplicitlySharedCampaignBudget

This field is not mutable on implicitly shared campaign budgets

CannotUpdateCampaignBudgetToImplicitlyShared

Cannot change explicitly shared campaign budgets back to implicitly shared ones.

CannotUpdateCampaignBudgetToExplicitlySharedWithoutName

An implicit campaign budget without a name cannot be changed to explicitly shared campaign budget.

CannotUpdateCampaignBudgetToExplicitlyShared

Cannot change an implicitly shared campaign budget to an explicitly shared one.

CannotUseImplicitlySharedCampaignBudgetWithMultipleCampaigns

Only explicitly shared campaign budgets can be used with multiple campaigns.

DuplicateName

A campaign budget with this name already exists.

MoneyAmountInWrongCurrency

A money amount was not in the expected currency.

MoneyAmountLessThanCurrencyMinimumCpc

A money amount was less than the minimum CPC for currency.

MoneyAmountTooLarge

A money amount was greater than the maximum allowed.

NegativeMoneyAmount

A money amount was negative.

NonMultipleOfMinimumCurrencyUnit

A money amount was not a multiple of a minimum unit.

TotalBudgetAmountMustBeUnsetForBudgetPeriodDaily

Total budget amount must be unset when BudgetPeriod is DAILY.

InvalidPeriod

The period of the budget is not allowed.

CannotUseAcceleratedDeliveryMode

Cannot use accelerated delivery method on this budget.

Implementations

impl CampaignBudgetError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CampaignBudgetError.

pub fn from_i32(value: i32) -> Option<CampaignBudgetError>

Converts an i32 to a CampaignBudgetError, or None if value is not a valid variant.

impl CampaignBudgetError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.