Enum google_api_proto::google::ads::googleads::v13::errors::conversion_upload_error_enum::ConversionUploadError

#[repr(i32)]
pub enum ConversionUploadError {

    Unspecified = 0,
    Unknown = 1,
    TooManyConversionsInRequest = 2,
    UnparseableGclid = 3,
    ConversionPrecedesEvent = 42,
    ExpiredEvent = 43,
    TooRecentEvent = 44,
    EventNotFound = 45,
    UnauthorizedCustomer = 8,
    InvalidConversionAction = 9,
    TooRecentConversionAction = 10,
    ConversionTrackingNotEnabledAtImpressionTime = 11,
    ExternalAttributionDataSetForNonExternallyAttributedConversionAction = 12,
    ExternalAttributionDataNotSetForExternallyAttributedConversionAction = 13,
    OrderIdNotPermittedForExternallyAttributedConversionAction = 14,
    OrderIdAlreadyInUse = 15,
    DuplicateOrderId = 16,
    TooRecentCall = 17,
    ExpiredCall = 18,
    CallNotFound = 19,
    ConversionPrecedesCall = 20,
    ConversionTrackingNotEnabledAtCallTime = 21,
    UnparseableCallersPhoneNumber = 22,
    ClickConversionAlreadyExists = 23,
    CallConversionAlreadyExists = 24,
    DuplicateClickConversionInRequest = 25,
    DuplicateCallConversionInRequest = 26,
    CustomVariableNotEnabled = 28,
    CustomVariableValueContainsPii = 29,
    InvalidCustomerForClick = 30,
    InvalidCustomerForCall = 31,
    ConversionNotCompliantWithAttPolicy = 32,
    ClickNotFound = 33,
    InvalidUserIdentifier = 34,
    ExternallyAttributedConversionActionNotPermittedWithUserIdentifier = 35,
    UnsupportedUserIdentifier = 36,
    GbraidWbraidBothSet = 38,
    UnparseableWbraid = 39,
    UnparseableGbraid = 40,
    OnePerClickConversionActionNotPermittedWithBraid = 46,
    CustomerDataPolicyProhibitsEnhancedConversions = 47,
    CustomerNotAcceptedCustomerDataTerms = 48,
    OrderIdContainsPii = 49,
    CustomerNotEnabledEnhancedConversionsForLeads = 50,

}

Enum describing possible conversion upload errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

TooManyConversionsInRequest = 2

The request contained more than 2000 conversions.

UnparseableGclid = 3

The specified gclid could not be decoded.

ConversionPrecedesEvent = 42

The specified conversion_date_time is before the event time associated with the given identifier or iOS URL parameter.

ExpiredEvent = 43

The click associated with the given identifier or iOS URL parameter is either too old to be imported or occurred outside of the click through lookback window for the specified conversion action.

TooRecentEvent = 44

The click associated with the given identifier or iOS URL parameter occurred too recently. Try uploading again after 6 hours have passed since the click occurred.

EventNotFound = 45

The click associated with the given identifier or iOS URL parameter could not be found in the system. This can happen if the identifier or iOS URL parameter are collected for non Google Ads clicks.

UnauthorizedCustomer = 8

The click associated with the given identifier or iOS URL parameter is owned by a customer account that the uploading customer does not manage.

InvalidConversionAction = 9

No upload eligible conversion action that matches the provided information can be found for the customer.

TooRecentConversionAction = 10

The specified conversion action was created too recently. Try the upload again after 4-6 hours have passed since the conversion action was created.

ConversionTrackingNotEnabledAtImpressionTime = 11

The click associated with the given identifier does not contain conversion tracking information.

ExternalAttributionDataSetForNonExternallyAttributedConversionAction = 12

The specified conversion action does not use an external attribution model, but external_attribution_data was set.

ExternalAttributionDataNotSetForExternallyAttributedConversionAction = 13

The specified conversion action uses an external attribution model, but external_attribution_data or one of its contained fields was not set. Both external_attribution_credit and external_attribution_model must be set for externally attributed conversion actions.

OrderIdNotPermittedForExternallyAttributedConversionAction = 14

Order IDs are not supported for conversion actions which use an external attribution model.

OrderIdAlreadyInUse = 15

A conversion with the same order id and conversion action combination already exists in our system.

DuplicateOrderId = 16

The request contained two or more conversions with the same order id and conversion action combination.

TooRecentCall = 17

The call occurred too recently. Try uploading again after 12 hours have passed since the call occurred.

ExpiredCall = 18

The click that initiated the call is too old for this conversion to be imported.

CallNotFound = 19

The call or the click leading to the call was not found.

ConversionPrecedesCall = 20

The specified conversion_date_time is before the call_start_date_time.

ConversionTrackingNotEnabledAtCallTime = 21

The click associated with the call does not contain conversion tracking information.

UnparseableCallersPhoneNumber = 22

The callerโs phone number cannot be parsed. It should be formatted either as E.164 โ+16502531234โ, International โ+64 3-331 6005โ or US national number โ6502531234โ.

ClickConversionAlreadyExists = 23

A conversion with this timestamp already exists for this click. To upload another conversion, use a different timestamp.

CallConversionAlreadyExists = 24

A conversion with this timestamp already exists for this call. To upload another conversion, use a different timestamp.

DuplicateClickConversionInRequest = 25

This conversion has the same click and timestamp as another conversion in the request. To upload another conversion for this click, use a different timestamp.

DuplicateCallConversionInRequest = 26

This conversion has the same call and timestamp as another conversion in the request. To upload another conversion for this call, use a different timestamp.

CustomVariableNotEnabled = 28

The custom variable is not enabled.

CustomVariableValueContainsPii = 29

The value of the custom variable contains personally identifiable information (PII), such as an email address or phone number.

InvalidCustomerForClick = 30

The click associated with the given identifier or iOS URL parameter isnโt from the account where conversion tracking is set up.

InvalidCustomerForCall = 31

The click associated with the given call isnโt from the account where conversion tracking is set up.

ConversionNotCompliantWithAttPolicy = 32

The conversion canโt be uploaded because the conversion source didnโt comply with the App Tracking Transparency (ATT) policy or the person who converted didnโt consent to tracking.

ClickNotFound = 33

No click was found for the provided user identifiers. This may be because the conversion did not come from a Google Ads campaign.

InvalidUserIdentifier = 34

The provided user identifier is not a SHA-256 hash. It is either unhashed or hashed using a different hash function.

ExternallyAttributedConversionActionNotPermittedWithUserIdentifier = 35

Conversion actions which use an external attribution model cannot be used with UserIdentifier.

UnsupportedUserIdentifier = 36

The provided user identifier is not supported. ConversionUploadService only supports hashed_email and hashed_phone_number.

GbraidWbraidBothSet = 38

gbraid and wbraid are both set in the request. Only one is allowed.

UnparseableWbraid = 39

The specified wbraid could not be decoded.

UnparseableGbraid = 40

The specified gbraid could not be decoded.

OnePerClickConversionActionNotPermittedWithBraid = 46

Conversion actions which use the one-per-click counting type cannot be used with gbraid or wbraid.

CustomerDataPolicyProhibitsEnhancedConversions = 47

Per our customer data policies, enhanced conversions have been prohibited in your account. If you have any questions, contact your Google representative.

CustomerNotAcceptedCustomerDataTerms = 48

The customer has not accepted the customer data terms in the conversion settings page.

OrderIdContainsPii = 49

The order_id contains personally identifiable information (PII), such as an email address or phone number.

CustomerNotEnabledEnhancedConversionsForLeads = 50

The customer has not enabled enhanced conversions for leads in the conversion settings page.

Implementations

impl ConversionUploadError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionUploadError.

pub fn from_i32(value: i32) -> Option<ConversionUploadError>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ConversionUploadError, or None if value is not a valid variant.

impl ConversionUploadError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.