Module google_api_proto::google::ads::googleads::v13::resources::listing_group_filter_dimension

Nested message and enum types in ListingGroupFilterDimension.

Structs

ProductBiddingCategory

One element of a bidding category at a certain level. Top-level categories are at level 1, their children at level 2, and so on. We currently support up to 5 levels. The user must specify a dimension type that indicates the level of the category. All cases of the same subdivision must have the same dimension type (category level).

ProductBrand

Brand of the product.

ProductChannel

Locality of a product offer.

ProductCondition

Condition of a product offer.

ProductCustomAttribute

Custom attribute of a product offer.

ProductItemId

Item id of a product offer.

ProductType

Type of a product offer.

Enums

Dimension

Dimension of one of the types below is always present.