Struct google_api_proto::google::ads::googleads::v13::resources::Customer

pub struct Customer {

    pub resource_name: String,
    pub id: Option<i64>,
    pub descriptive_name: Option<String>,
    pub currency_code: Option<String>,
    pub time_zone: Option<String>,
    pub tracking_url_template: Option<String>,
    pub final_url_suffix: Option<String>,
    pub auto_tagging_enabled: Option<bool>,
    pub has_partners_badge: Option<bool>,
    pub manager: Option<bool>,
    pub test_account: Option<bool>,
    pub call_reporting_setting: Option<CallReportingSetting>,
    pub conversion_tracking_setting: Option<ConversionTrackingSetting>,
    pub remarketing_setting: Option<RemarketingSetting>,
    pub pay_per_conversion_eligibility_failure_reasons: Vec<i32>,
    pub optimization_score: Option<f64>,
    pub optimization_score_weight: f64,
    pub status: i32,
    pub location_asset_auto_migration_done: Option<bool>,
    pub image_asset_auto_migration_done: Option<bool>,
    pub location_asset_auto_migration_done_date_time: Option<String>,
    pub image_asset_auto_migration_done_date_time: Option<String>,

}

A customer.

Fields
resource_name: String

Immutable. The resource name of the customer. Customer resource names have the form:

customers/{customer_id}

id: Option<i64>

Output only. The ID of the customer.

descriptive_name: Option<String>

Optional, non-unique descriptive name of the customer.

currency_code: Option<String>

Immutable. The currency in which the account operates. A subset of the currency codes from the ISO 4217 standard is supported.

time_zone: Option<String>

Immutable. The local timezone ID of the customer.

tracking_url_template: Option<String>

The URL template for constructing a tracking URL out of parameters. Only mutable in an update operation.

final_url_suffix: Option<String>

The URL template for appending params to the final URL. Only mutable in an update operation.

auto_tagging_enabled: Option<bool>

Whether auto-tagging is enabled for the customer.

has_partners_badge: Option<bool>

Output only. Whether the Customer has a Partners program badge. If the Customer is not associated with the Partners program, this will be false. For more information, see https://support.google.com/partners/answer/3125774.

manager: Option<bool>

Output only. Whether the customer is a manager.

test_account: Option<bool>

Output only. Whether the customer is a test account.

call_reporting_setting: Option<CallReportingSetting>

Call reporting setting for a customer. Only mutable in an update operation.

conversion_tracking_setting: Option<ConversionTrackingSetting>

Output only. Conversion tracking setting for a customer.

remarketing_setting: Option<RemarketingSetting>

Output only. Remarketing setting for a customer.

pay_per_conversion_eligibility_failure_reasons: Vec<i32>

Output only. Reasons why the customer is not eligible to use PaymentMode.CONVERSIONS. If the list is empty, the customer is eligible. This field is read-only.

optimization_score: Option<f64>

Output only. Optimization score of the customer.

Optimization score is an estimate of how well a customer’s campaigns are set to perform. It ranges from 0% (0.0) to 100% (1.0). This field is null for all manager customers, and for unscored non-manager customers.

See “About optimization score” at https://support.google.com/google-ads/answer/9061546.

This field is read-only.

optimization_score_weight: f64

Output only. Optimization score weight of the customer.

Optimization score weight can be used to compare/aggregate optimization scores across multiple non-manager customers. The aggregate optimization score of a manager is computed as the sum over all of their customers of Customer.optimization_score * Customer.optimization_score_weight. This field is 0 for all manager customers, and for unscored non-manager customers.

This field is read-only.

status: i32

Output only. The status of the customer.

location_asset_auto_migration_done: Option<bool>

Output only. True if feed based location has been migrated to asset based location.

image_asset_auto_migration_done: Option<bool>

Output only. True if feed based image has been migrated to asset based image.

location_asset_auto_migration_done_date_time: Option<String>

Output only. Timestamp of migration from feed based location to asset base location in yyyy-MM-dd HH:mm:ss format.

image_asset_auto_migration_done_date_time: Option<String>

Output only. Timestamp of migration from feed based image to asset base image in yyyy-MM-dd HH:mm:ss format.

Implementations

impl Customer

pub fn pay_per_conversion_eligibility_failure_reasons( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<CustomerPayPerConversionEligibilityFailureReason>>

Returns an iterator which yields the valid enum values contained in pay_per_conversion_eligibility_failure_reasons.

pub fn push_pay_per_conversion_eligibility_failure_reasons( &mut self, value: CustomerPayPerConversionEligibilityFailureReason )

Appends the provided enum value to pay_per_conversion_eligibility_failure_reasons.

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn descriptive_name(&self) -> &str

Returns the value of descriptive_name, or the default value if descriptive_name is unset.

pub fn currency_code(&self) -> &str

Returns the value of currency_code, or the default value if currency_code is unset.

pub fn time_zone(&self) -> &str

Returns the value of time_zone, or the default value if time_zone is unset.

pub fn tracking_url_template(&self) -> &str

Returns the value of tracking_url_template, or the default value if tracking_url_template is unset.

pub fn final_url_suffix(&self) -> &str

Returns the value of final_url_suffix, or the default value if final_url_suffix is unset.

pub fn auto_tagging_enabled(&self) -> bool

Returns the value of auto_tagging_enabled, or the default value if auto_tagging_enabled is unset.

pub fn has_partners_badge(&self) -> bool

Returns the value of has_partners_badge, or the default value if has_partners_badge is unset.

pub fn manager(&self) -> bool

Returns the value of manager, or the default value if manager is unset.

pub fn test_account(&self) -> bool

Returns the value of test_account, or the default value if test_account is unset.

pub fn optimization_score(&self) -> f64

Returns the value of optimization_score, or the default value if optimization_score is unset.

pub fn status(&self) -> CustomerStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: CustomerStatus)

Sets status to the provided enum value.

pub fn location_asset_auto_migration_done(&self) -> bool

Returns the value of location_asset_auto_migration_done, or the default value if location_asset_auto_migration_done is unset.

pub fn image_asset_auto_migration_done(&self) -> bool

Returns the value of image_asset_auto_migration_done, or the default value if image_asset_auto_migration_done is unset.

pub fn location_asset_auto_migration_done_date_time(&self) -> &str

Returns the value of location_asset_auto_migration_done_date_time, or the default value if location_asset_auto_migration_done_date_time is unset.

pub fn image_asset_auto_migration_done_date_time(&self) -> &str

Returns the value of image_asset_auto_migration_done_date_time, or the default value if image_asset_auto_migration_done_date_time is unset.