Enum google_api_proto::google::ads::googleads::v14::enums::campaign_primary_status_reason_enum::CampaignPrimaryStatusReason

#[repr(i32)]
pub enum CampaignPrimaryStatusReason {

    Unspecified = 0,
    Unknown = 1,
    CampaignRemoved = 2,
    CampaignPaused = 3,
    CampaignPending = 4,
    CampaignEnded = 5,
    CampaignDraft = 6,
    BiddingStrategyMisconfigured = 7,
    BiddingStrategyLimited = 8,
    BiddingStrategyLearning = 9,
    BiddingStrategyConstrained = 10,
    BudgetConstrained = 11,
    BudgetMisconfigured = 12,
    SearchVolumeLimited = 13,
    AdGroupsPaused = 14,
    NoAdGroups = 15,
    KeywordsPaused = 16,
    NoKeywords = 17,
    AdGroupAdsPaused = 18,
    NoAdGroupAds = 19,
    HasAdsLimitedByPolicy = 20,
    HasAdsDisapproved = 21,
    MostAdsUnderReview = 22,
    MissingLeadFormExtension = 23,
    MissingCallExtension = 24,
    LeadFormExtensionUnderReview = 25,
    LeadFormExtensionDisapproved = 26,
    CallExtensionUnderReview = 27,
    CallExtensionDisapproved = 28,
    NoMobileApplicationAdGroupCriteria = 29,
    CampaignGroupPaused = 30,
    CampaignGroupAllGroupBudgetsEnded = 31,
    AppNotReleased = 32,
    AppPartiallyReleased = 33,
    HasAssetGroupsDisapproved = 34,
    HasAssetGroupsLimitedByPolicy = 35,
    MostAssetGroupsUnderReview = 36,

}

Enum describing the possible campaign primary status reasons. Provides insight into why a campaign is not serving or not serving optimally. These reasons are aggregated to determine an overall campaign primary status.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

CampaignRemoved = 2

The user-specified campaign status is removed.

CampaignPaused = 3

The user-specified campaign status is paused.

CampaignPending = 4

The user-specified time for this campaign to start is in the future.

CampaignEnded = 5

The user-specified time for this campaign to end has passed.

CampaignDraft = 6

The campaign is a draft.

BiddingStrategyMisconfigured = 7

The bidding strategy has incorrect user-specified settings.

BiddingStrategyLimited = 8

The bidding strategy is limited by user-specified settings such as lack of data or similar.

BiddingStrategyLearning = 9

The automated bidding system is adjusting to user-specified changes to the bidding strategy.

BiddingStrategyConstrained = 10

Campaign could capture more conversion value by adjusting CPA/ROAS targets.

BudgetConstrained = 11

The budget is limiting the campaignโs ability to serve.

BudgetMisconfigured = 12

The budget has incorrect user-specified settings.

SearchVolumeLimited = 13

Campaign is not targeting all relevant queries.

AdGroupsPaused = 14

The user-specified ad group statuses are all paused.

NoAdGroups = 15

No eligible ad groups exist in this campaign.

KeywordsPaused = 16

The user-specified keyword statuses are all paused.

NoKeywords = 17

No eligible keywords exist in this campaign.

AdGroupAdsPaused = 18

The user-specified ad group ad statuses are all paused.

NoAdGroupAds = 19

No eligible ad group ads exist in this campaign.

HasAdsLimitedByPolicy = 20

At least one ad in this campaign is limited by policy.

HasAdsDisapproved = 21

At least one ad in this campaign is disapproved.

MostAdsUnderReview = 22

Most ads in this campaign are pending review.

MissingLeadFormExtension = 23

The campaign has a lead form goal, and the lead form extension is missing.

MissingCallExtension = 24

The campaign has a call goal, and the call extension is missing.

LeadFormExtensionUnderReview = 25

The lead form extension is under review.

LeadFormExtensionDisapproved = 26

The lead extension is disapproved.

CallExtensionUnderReview = 27

The call extension is under review.

CallExtensionDisapproved = 28

The call extension is disapproved.

NoMobileApplicationAdGroupCriteria = 29

No eligible mobile application ad group criteria exist in this campaign.

CampaignGroupPaused = 30

The user-specified campaign group status is paused.

CampaignGroupAllGroupBudgetsEnded = 31

The user-specified times of all group budgets associated with the parent campaign group has passed.

AppNotReleased = 32

The app associated with this ACi campaign is not released in the target countries of the campaign.

AppPartiallyReleased = 33

The app associated with this ACi campaign is partially released in the target countries of the campaign.

HasAssetGroupsDisapproved = 34

At least one asset group in this campaign is disapproved.

HasAssetGroupsLimitedByPolicy = 35

At least one asset group in this campaign is limited by policy.

MostAssetGroupsUnderReview = 36

Most asset groups in this campaign are pending review.

Implementations

impl CampaignPrimaryStatusReason

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CampaignPrimaryStatusReason.

pub fn from_i32(value: i32) -> Option<CampaignPrimaryStatusReason>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a CampaignPrimaryStatusReason, or None if value is not a valid variant.

impl CampaignPrimaryStatusReason

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.