Enum google_api_proto::google::ads::googleads::v14::enums::conversion_action_type_enum::ConversionActionType

#[repr(i32)]
pub enum ConversionActionType {

    Unspecified = 0,
    Unknown = 1,
    AdCall = 2,
    ClickToCall = 3,
    GooglePlayDownload = 4,
    GooglePlayInAppPurchase = 5,
    UploadCalls = 6,
    UploadClicks = 7,
    Webpage = 8,
    WebsiteCall = 9,
    StoreSalesDirectUpload = 10,
    StoreSales = 11,
    FirebaseAndroidFirstOpen = 12,
    FirebaseAndroidInAppPurchase = 13,
    FirebaseAndroidCustom = 14,
    FirebaseIosFirstOpen = 15,
    FirebaseIosInAppPurchase = 16,
    FirebaseIosCustom = 17,
    ThirdPartyAppAnalyticsAndroidFirstOpen = 18,
    ThirdPartyAppAnalyticsAndroidInAppPurchase = 19,
    ThirdPartyAppAnalyticsAndroidCustom = 20,
    ThirdPartyAppAnalyticsIosFirstOpen = 21,
    ThirdPartyAppAnalyticsIosInAppPurchase = 22,
    ThirdPartyAppAnalyticsIosCustom = 23,
    AndroidAppPreRegistration = 24,
    AndroidInstallsAllOtherApps = 25,
    FloodlightAction = 26,
    FloodlightTransaction = 27,
    GoogleHosted = 28,
    LeadFormSubmit = 29,
    Salesforce = 30,
    SearchAds360 = 31,
    SmartCampaignAdClicksToCall = 32,
    SmartCampaignMapClicksToCall = 33,
    SmartCampaignMapDirections = 34,
    SmartCampaignTrackedCalls = 35,
    StoreVisits = 36,
    WebpageCodeless = 37,
    UniversalAnalyticsGoal = 38,
    UniversalAnalyticsTransaction = 39,
    GoogleAnalytics4Custom = 40,
    GoogleAnalytics4Purchase = 41,

}

Possible types of a conversion action.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

AdCall = 2

Conversions that occur when a user clicks on an adโs call extension.

ClickToCall = 3

Conversions that occur when a user on a mobile device clicks a phone number.

GooglePlayDownload = 4

Conversions that occur when a user downloads a mobile app from the Google Play Store.

GooglePlayInAppPurchase = 5

Conversions that occur when a user makes a purchase in an app through Android billing.

UploadCalls = 6

Call conversions that are tracked by the advertiser and uploaded.

UploadClicks = 7

Conversions that are tracked by the advertiser and uploaded with attributed clicks.

Webpage = 8

Conversions that occur on a webpage.

WebsiteCall = 9

Conversions that occur when a user calls a dynamically-generated phone number from an advertiserโs website.

StoreSalesDirectUpload = 10

Store Sales conversion based on first-party or third-party merchant data uploads. Only customers on the allowlist can use store sales direct upload types.

StoreSales = 11

Store Sales conversion based on first-party or third-party merchant data uploads and/or from in-store purchases using cards from payment networks. Only customers on the allowlist can use store sales types. Read only.

FirebaseAndroidFirstOpen = 12

Android app first open conversions tracked through Firebase.

FirebaseAndroidInAppPurchase = 13

Android app in app purchase conversions tracked through Firebase.

FirebaseAndroidCustom = 14

Android app custom conversions tracked through Firebase.

FirebaseIosFirstOpen = 15

iOS app first open conversions tracked through Firebase.

FirebaseIosInAppPurchase = 16

iOS app in app purchase conversions tracked through Firebase.

FirebaseIosCustom = 17

iOS app custom conversions tracked through Firebase.

ThirdPartyAppAnalyticsAndroidFirstOpen = 18

Android app first open conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsAndroidInAppPurchase = 19

Android app in app purchase conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsAndroidCustom = 20

Android app custom conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosFirstOpen = 21

iOS app first open conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosInAppPurchase = 22

iOS app in app purchase conversions tracked through Third Party App Analytics.

ThirdPartyAppAnalyticsIosCustom = 23

iOS app custom conversions tracked through Third Party App Analytics.

AndroidAppPreRegistration = 24

Conversions that occur when a user pre-registers a mobile app from the Google Play Store. Read only.

AndroidInstallsAllOtherApps = 25

Conversions that track all Google Play downloads which arenโt tracked by an app-specific type. Read only.

FloodlightAction = 26

Floodlight activity that counts the number of times that users have visited a particular webpage after seeing or clicking on one of an advertiserโs ads. Read only.

FloodlightTransaction = 27

Floodlight activity that tracks the number of sales made or the number of items purchased. Can also capture the total value of each sale. Read only.

GoogleHosted = 28

Conversions that track local actions from Googleโs products and services after interacting with an ad. Read only.

LeadFormSubmit = 29

Conversions reported when a user submits a lead form. Read only.

Salesforce = 30

Conversions that come from Salesforce. Read only.

SearchAds360 = 31

Conversions imported from Search Ads 360 Floodlight data. Read only.

SmartCampaignAdClicksToCall = 32

Call conversions that occur on Smart campaign Ads without call tracking setup, using Smart campaign custom criteria. Read only.

SmartCampaignMapClicksToCall = 33

The user clicks on a call element within Google Maps. Smart campaign only. Read only.

SmartCampaignMapDirections = 34

The user requests directions to a business location within Google Maps. Smart campaign only. Read only.

SmartCampaignTrackedCalls = 35

Call conversions that occur on Smart campaign Ads with call tracking setup, using Smart campaign custom criteria. Read only.

StoreVisits = 36

Conversions that occur when a user visits an advertiserโs retail store. Read only.

WebpageCodeless = 37

Conversions created from website events (such as form submissions or page loads), that donโt use individually coded event snippets. Read only.

UniversalAnalyticsGoal = 38

Conversions that come from linked Universal Analytics goals.

UniversalAnalyticsTransaction = 39

Conversions that come from linked Universal Analytics transactions.

GoogleAnalytics4Custom = 40

Conversions that come from linked Google Analytics 4 custom event conversions.

GoogleAnalytics4Purchase = 41

Conversions that come from linked Google Analytics 4 purchase conversions.

Implementations

impl ConversionActionType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionActionType.

pub fn from_i32(value: i32) -> Option<ConversionActionType>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ConversionActionType, or None if value is not a valid variant.

impl ConversionActionType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.