Enum google_api_proto::google::ads::googleads::v14::enums::placeholder_type_enum::PlaceholderType

#[repr(i32)]
pub enum PlaceholderType {

    Unspecified,
    Unknown,
    Sitelink,
    Call,
    App,
    Location,
    AffiliateLocation,
    Callout,
    StructuredSnippet,
    Message,
    Price,
    Promotion,
    AdCustomizer,
    DynamicEducation,
    DynamicFlight,
    DynamicCustom,
    DynamicHotel,
    DynamicRealEstate,
    DynamicTravel,
    DynamicLocal,
    DynamicJob,
    Image,

}

Possible placeholder types for a feed mapping.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

Sitelink

Lets you show links in your ad to pages from your website, including the main landing page.

Call

Lets you attach a phone number to an ad, allowing customers to call directly from the ad.

App

Lets you provide users with a link that points to a mobile app in addition to a website.

Location

Lets you show locations of businesses from your Business Profile in your ad. This helps people find your locations by showing your ads with your address, a map to your location, or the distance to your business. This extension type is useful to draw customers to your brick-and-mortar location.

AffiliateLocation

If you sell your product through retail chains, affiliate location extensions let you show nearby stores that carry your products.

Callout

Lets you include additional text with your search ads that provide detailed information about your business, including products and services you offer. Callouts appear in ads at the top and bottom of Google search results.

StructuredSnippet

Lets you add more info to your ad, specific to some predefined categories such as types, brands, styles, etc. A minimum of 3 text (SNIPPETS) values are required.

Message

Allows users to see your ad, click an icon, and contact you directly by text message. With one tap on your ad, people can contact you to book an appointment, get a quote, ask for information, or request a service.

Price

Lets you display prices for a list of items along with your ads. A price feed is composed of three to eight price table rows.

Promotion

Lets you highlight sales and other promotions that let users see how they can save by buying now.

AdCustomizer

Lets you dynamically inject custom data into the title and description of your ads.

DynamicEducation

Indicates that this feed is for education dynamic remarketing.

DynamicFlight

Indicates that this feed is for flight dynamic remarketing.

DynamicCustom

Indicates that this feed is for a custom dynamic remarketing type. Use this only if the other business types donโt apply to your products or services.

DynamicHotel

Indicates that this feed is for hotels and rentals dynamic remarketing.

DynamicRealEstate

Indicates that this feed is for real estate dynamic remarketing.

DynamicTravel

Indicates that this feed is for travel dynamic remarketing.

DynamicLocal

Indicates that this feed is for local deals dynamic remarketing.

DynamicJob

Indicates that this feed is for job dynamic remarketing.

Image

Lets you attach an image to an ad.

Implementations

impl PlaceholderType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of PlaceholderType.

pub fn from_i32(value: i32) -> Option<PlaceholderType>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a PlaceholderType, or None if value is not a valid variant.

impl PlaceholderType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.