Enum google_api_proto::google::ads::googleads::v14::errors::campaign_error_enum::CampaignError

#[repr(i32)]
pub enum CampaignError {

    Unspecified = 0,
    Unknown = 1,
    CannotTargetContentNetwork = 3,
    CannotTargetSearchNetwork = 4,
    CannotTargetSearchNetworkWithoutGoogleSearch = 5,
    CannotTargetGoogleSearchForCpmCampaign = 6,
    CampaignMustTargetAtLeastOneNetwork = 7,
    CannotTargetPartnerSearchNetwork = 8,
    CannotTargetContentNetworkOnlyWithCriteriaLevelBiddingStrategy = 9,
    CampaignDurationMustContainAllRunnableTrials = 10,
    CannotModifyForTrialCampaign = 11,
    DuplicateCampaignName = 12,
    IncompatibleCampaignField = 13,
    InvalidCampaignName = 14,
    InvalidAdServingOptimizationStatus = 15,
    InvalidTrackingUrl = 16,
    CannotSetBothTrackingUrlTemplateAndTrackingSetting = 17,
    MaxImpressionsNotInRange = 18,
    TimeUnitNotSupported = 19,
    InvalidOperationIfServingStatusHasEnded = 20,
    BudgetCannotBeShared = 21,
    CampaignCannotUseSharedBudget = 22,
    CannotChangeBudgetOnCampaignWithTrials = 23,
    CampaignLabelDoesNotExist = 24,
    CampaignLabelAlreadyExists = 25,
    MissingShoppingSetting = 26,
    InvalidShoppingSalesCountry = 27,
    AdvertisingChannelTypeNotAvailableForAccountType = 31,
    InvalidAdvertisingChannelSubType = 32,
    AtLeastOneConversionMustBeSelected = 33,
    CannotSetAdRotationMode = 34,
    CannotModifyStartDateIfAlreadyStarted = 35,
    CannotSetDateToPast = 36,
    MissingHotelCustomerLink = 37,
    InvalidHotelCustomerLink = 38,
    MissingHotelSetting = 39,
    CannotUseSharedCampaignBudgetWhilePartOfCampaignGroup = 40,
    AppNotFound = 41,
    ShoppingEnableLocalNotSupportedForCampaignType = 42,
    MerchantNotAllowedForComparisonListingAds = 43,
    InsufficientAppInstallsCount = 44,
    SensitiveCategoryApp = 45,
    HecAgreementRequired = 46,
    NotCompatibleWithViewThroughConversionOptimization = 49,
    InvalidExcludedParentAssetFieldType = 48,
    CannotCreateAppPreRegistrationForNonAndroidApp = 50,
    AppNotAvailableToCreateAppPreRegistrationCampaign = 51,
    IncompatibleBudgetType = 52,
    LocalServicesDuplicateCategoryBid = 53,
    LocalServicesInvalidCategoryBid = 54,
    LocalServicesMissingCategoryBid = 55,
    InvalidStatusChange = 57,
    MissingTravelCustomerLink = 58,
    InvalidTravelCustomerLink = 59,
    InvalidExcludedParentAssetSetType = 62,
    AssetSetNotAHotelPropertyAssetSet = 63,
    HotelPropertyAssetSetOnlyForPerformanceMaxForTravelGoals = 64,
    AverageDailySpendTooHigh = 65,
    CannotAttachToRemovedCampaignGroup = 66,
    CannotAttachToBiddingStrategy = 67,
    CannotChangeBudgetPeriod = 68,
    NotEnoughConversions = 71,
    CannotSetMoreThanOneConversionAction = 72,
    NotCompatibleWithBudgetType = 73,
    NotCompatibleWithUploadClicksConversion = 74,
    AppIdMustMatchConversionActionAppId = 76,
    ConversionActionWithDownloadCategoryNotAllowed = 77,
    ConversionActionWithDownloadCategoryRequired = 78,
    ConversionTrackingNotEnabled = 79,
    NotCompatibleWithBiddingStrategyType = 80,
    NotCompatibleWithGoogleAttributionConversions = 81,
    ConversionLagTooHigh = 82,
    NotLinkedAdvertisingPartner = 83,
    InvalidNumberOfAdvertisingPartnerIds = 84,

}

Enum describing possible campaign errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

CannotTargetContentNetwork = 3

Cannot target content network.

CannotTargetSearchNetwork = 4

Cannot target search network.

CannotTargetSearchNetworkWithoutGoogleSearch = 5

Cannot cover search network without google search network.

CannotTargetGoogleSearchForCpmCampaign = 6

Cannot target Google Search network for a CPM campaign.

CampaignMustTargetAtLeastOneNetwork = 7

Must target at least one network.

CannotTargetPartnerSearchNetwork = 8

Only some Google partners are allowed to target partner search network.

CannotTargetContentNetworkOnlyWithCriteriaLevelBiddingStrategy = 9

Cannot target content network only as campaign has criteria-level bidding strategy.

CampaignDurationMustContainAllRunnableTrials = 10

Cannot modify the start or end date such that the campaign duration would not contain the durations of all runnable trials.

CannotModifyForTrialCampaign = 11

Cannot modify dates, budget or status of a trial campaign.

DuplicateCampaignName = 12

Trying to modify the name of an active or paused campaign, where the name is already assigned to another active or paused campaign.

IncompatibleCampaignField = 13

Two fields are in conflicting modes.

InvalidCampaignName = 14

Campaign name cannot be used.

InvalidAdServingOptimizationStatus = 15

Given status is invalid.

InvalidTrackingUrl = 16

Error in the campaign level tracking URL.

CannotSetBothTrackingUrlTemplateAndTrackingSetting = 17

Cannot set both tracking URL template and tracking setting. A user has to clear legacy tracking setting in order to add tracking URL template.

MaxImpressionsNotInRange = 18

The maximum number of impressions for Frequency Cap should be an integer greater than 0.

TimeUnitNotSupported = 19

Only the Day, Week and Month time units are supported.

InvalidOperationIfServingStatusHasEnded = 20

Operation not allowed on a campaign whose serving status has ended

BudgetCannotBeShared = 21

This budget is exclusively linked to a Campaign that is using experiments so it cannot be shared.

CampaignCannotUseSharedBudget = 22

Campaigns using experiments cannot use a shared budget.

CannotChangeBudgetOnCampaignWithTrials = 23

A different budget cannot be assigned to a campaign when there are running or scheduled trials.

CampaignLabelDoesNotExist = 24

No link found between the campaign and the label.

CampaignLabelAlreadyExists = 25

The label has already been attached to the campaign.

MissingShoppingSetting = 26

A ShoppingSetting was not found when creating a shopping campaign.

InvalidShoppingSalesCountry = 27

The country in shopping setting is not an allowed country.

AdvertisingChannelTypeNotAvailableForAccountType = 31

The requested channel type is not available according to the customerโs account setting.

InvalidAdvertisingChannelSubType = 32

The AdvertisingChannelSubType is not a valid subtype of the primary channel type.

AtLeastOneConversionMustBeSelected = 33

At least one conversion must be selected.

CannotSetAdRotationMode = 34

Setting ad rotation mode for a campaign is not allowed. Ad rotation mode at campaign is deprecated.

CannotModifyStartDateIfAlreadyStarted = 35

Trying to change start date on a campaign that has started.

CannotSetDateToPast = 36

Trying to modify a date into the past.

MissingHotelCustomerLink = 37

Hotel center id in the hotel setting does not match any customer links.

InvalidHotelCustomerLink = 38

Hotel center id in the hotel setting must match an active customer link.

MissingHotelSetting = 39

Hotel setting was not found when creating a hotel ads campaign.

CannotUseSharedCampaignBudgetWhilePartOfCampaignGroup = 40

A Campaign cannot use shared campaign budgets and be part of a campaign group.

AppNotFound = 41

The app ID was not found.

ShoppingEnableLocalNotSupportedForCampaignType = 42

Campaign.shopping_setting.enable_local is not supported for the specified campaign type.

MerchantNotAllowedForComparisonListingAds = 43

The merchant does not support the creation of campaigns for Shopping Comparison Listing Ads.

InsufficientAppInstallsCount = 44

The App campaign for engagement cannot be created because there arenโt enough installs.

SensitiveCategoryApp = 45

The App campaign for engagement cannot be created because the app is sensitive.

HecAgreementRequired = 46

Customers with Housing, Employment, or Credit ads must accept updated personalized ads policy to continue creating campaigns.

NotCompatibleWithViewThroughConversionOptimization = 49

The field is not compatible with view through conversion optimization.

InvalidExcludedParentAssetFieldType = 48

The field type cannot be excluded because an active campaign-asset link of this type exists.

CannotCreateAppPreRegistrationForNonAndroidApp = 50

The app pre-registration campaign cannot be created for non-Android applications.

AppNotAvailableToCreateAppPreRegistrationCampaign = 51

The campaign cannot be created since the app is not available for pre-registration in any country.

IncompatibleBudgetType = 52

The type of the Budget is not compatible with this Campaign.

LocalServicesDuplicateCategoryBid = 53

Category bid list in the local services campaign setting contains multiple bids for the same category ID.

LocalServicesInvalidCategoryBid = 54

Category bid list in the local services campaign setting contains a bid for an invalid category ID.

LocalServicesMissingCategoryBid = 55

Category bid list in the local services campaign setting is missing a bid for a category ID that must be present.

InvalidStatusChange = 57

The requested change in status is not supported.

MissingTravelCustomerLink = 58

Travel Campaignโs travel_account_id does not match any customer links.

InvalidTravelCustomerLink = 59

Travel Campaignโs travel_account_id matches an existing customer link but the customer link is not active.

InvalidExcludedParentAssetSetType = 62

The asset set type is invalid to be set in excluded_parent_asset_set_types field.

AssetSetNotAHotelPropertyAssetSet = 63

Campaign.hotel_property_asset_set must point to an asset set of type HOTEL_PROPERTY.

HotelPropertyAssetSetOnlyForPerformanceMaxForTravelGoals = 64

The hotel property asset set can only be set on Performance Max for travel goals campaigns.

AverageDailySpendTooHigh = 65

Customerโs average daily spend is too high to enable this feature.

CannotAttachToRemovedCampaignGroup = 66

Cannot attach the campaign to a deleted campaign group.

CannotAttachToBiddingStrategy = 67

Cannot attach the campaign to this bidding strategy.

CannotChangeBudgetPeriod = 68

A budget with a different period cannot be assigned to the campaign.

NotEnoughConversions = 71

Customer does not have enough conversions to enable this feature.

CannotSetMoreThanOneConversionAction = 72

This campaign type can only have one conversion action.

NotCompatibleWithBudgetType = 73

The field is not compatible with the budget type.

NotCompatibleWithUploadClicksConversion = 74

The feature is incompatible with ConversionActionType.UPLOAD_CLICKS.

AppIdMustMatchConversionActionAppId = 76

App campaign setting app ID must match selective optimization conversion action app ID.

ConversionActionWithDownloadCategoryNotAllowed = 77

Selective optimization conversion action with Download category is not allowed.

ConversionActionWithDownloadCategoryRequired = 78

One software download for selective optimization conversion action is required for this campaign conversion action.

ConversionTrackingNotEnabled = 79

Conversion tracking is not enabled and is required for this feature.

NotCompatibleWithBiddingStrategyType = 80

The field is not compatible with the bidding strategy type.

NotCompatibleWithGoogleAttributionConversions = 81

Campaign is not compatible with a conversion tracker that has Google attribution enabled.

ConversionLagTooHigh = 82

Customer level conversion lag is too high.

NotLinkedAdvertisingPartner = 83

The advertiser set as an advertising partner is not an actively linked advertiser to this customer.

InvalidNumberOfAdvertisingPartnerIds = 84

Invalid number of advertising partner IDs.

Implementations

impl CampaignError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CampaignError.

pub fn from_i32(value: i32) -> Option<CampaignError>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a CampaignError, or None if value is not a valid variant.

impl CampaignError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.