Enum google_api_proto::google::ads::googleads::v14::errors::conversion_upload_error_enum::ConversionUploadError

#[repr(i32)]
pub enum ConversionUploadError {

    Unspecified,
    Unknown,
    TooManyConversionsInRequest,
    UnparseableGclid,
    ConversionPrecedesEvent,
    ExpiredEvent,
    TooRecentEvent,
    EventNotFound,
    UnauthorizedCustomer,
    InvalidConversionAction,
    TooRecentConversionAction,
    ConversionTrackingNotEnabledAtImpressionTime,
    ExternalAttributionDataSetForNonExternallyAttributedConversionAction,
    ExternalAttributionDataNotSetForExternallyAttributedConversionAction,
    OrderIdNotPermittedForExternallyAttributedConversionAction,
    OrderIdAlreadyInUse,
    DuplicateOrderId,
    TooRecentCall,
    ExpiredCall,
    CallNotFound,
    ConversionPrecedesCall,
    ConversionTrackingNotEnabledAtCallTime,
    UnparseableCallersPhoneNumber,
    ClickConversionAlreadyExists,
    CallConversionAlreadyExists,
    DuplicateClickConversionInRequest,
    DuplicateCallConversionInRequest,
    CustomVariableNotEnabled,
    CustomVariableValueContainsPii,
    InvalidCustomerForClick,
    InvalidCustomerForCall,
    ConversionNotCompliantWithAttPolicy,
    ClickNotFound,
    InvalidUserIdentifier,
    ExternallyAttributedConversionActionNotPermittedWithUserIdentifier,
    UnsupportedUserIdentifier,
    GbraidWbraidBothSet,
    UnparseableWbraid,
    UnparseableGbraid,
    OnePerClickConversionActionNotPermittedWithBraid,
    CustomerDataPolicyProhibitsEnhancedConversions,
    CustomerNotAcceptedCustomerDataTerms,
    OrderIdContainsPii,
    CustomerNotEnabledEnhancedConversionsForLeads,

}

Enum describing possible conversion upload errors.

Variants

Unspecified

Enum unspecified.

Unknown

Used for return value only. Represents value unknown in this version.

TooManyConversionsInRequest

Upload fewer than 2001 events in a single request.

UnparseableGclid

The imported gclid could not be decoded. Make sure you have not modified the click IDs.

ConversionPrecedesEvent

The imported event has a conversion_date_time that precedes the click. Make sure your conversion_date_time is correct and try again.

ExpiredEvent

The imported event can’t be recorded because its click occurred before this conversion’s click-through window. Make sure you import the most recent data.

TooRecentEvent

The click associated with the given identifier or iOS URL parameter occurred less than 6 hours ago. Retry after 6 hours have passed.

EventNotFound

The imported event could not be attributed to a click. This may be because the event did not come from a Google Ads campaign.

UnauthorizedCustomer

The click ID or call is associated with an Ads account you don’t have access to. Make sure you import conversions for accounts managed by your manager account.

InvalidConversionAction

Make sure you specify a valid conversion action set up for offline import.

TooRecentConversionAction

Can’t import events to a conversion action that was just created. Try importing again in 6 hours.

ConversionTrackingNotEnabledAtImpressionTime

At the time of the click, conversion tracking was not enabled in the effective conversion account of the click’s Google Ads account.

ExternalAttributionDataSetForNonExternallyAttributedConversionAction

The imported event includes external attribution data, but the conversion action isn’t set up to use an external attribution model. Make sure the conversion action is correctly configured and try again.

ExternalAttributionDataNotSetForExternallyAttributedConversionAction

The conversion action is set up to use an external attribution model, but the imported event is missing data. Make sure imported events include the external attribution credit and all necessary fields.

OrderIdNotPermittedForExternallyAttributedConversionAction

Order IDs can’t be used for a conversion measured with an external attribution model. Make sure the conversion is correctly configured and imported events include only necessary data and try again.

OrderIdAlreadyInUse

The imported event includes an order ID that was previously recorded, so the event was not processed.

DuplicateOrderId

Imported events include multiple conversions with the same order ID and were not processed. Make sure order IDs are unique and try again.

TooRecentCall

Can’t import calls that occurred less than 6 hours ago. Try uploading again in 6 hours.

ExpiredCall

The call can’t be recorded because it occurred before this conversion action’s lookback window. Make sure your import is configured to get the most recent data.

CallNotFound

The call or click leading to the imported event can’t be found. Make sure your data source is set up to include correct identifiers.

ConversionPrecedesCall

The call has a conversion_date_time that precedes the associated click. Make sure your conversion_date_time is correct.

ConversionTrackingNotEnabledAtCallTime

At the time of the imported call, conversion tracking was not enabled in the effective conversion account of the click’s Google Ads account.

UnparseableCallersPhoneNumber

Make sure phone numbers are formatted as E.164 (+16502531234), International (+64 3-331 6005), or US national number (6502531234).

ClickConversionAlreadyExists

The imported event has the same click and conversion_date_time as an existing conversion. Use a unique conversion_date_time or order ID for each unique event and try again.

CallConversionAlreadyExists

The imported call has the same conversion_date_time as an existing conversion. Make sure your conversion_date_time correctly configured and try again.

DuplicateClickConversionInRequest

Multiple events have the same click and conversion_date_time. Make sure your conversion_date_time is correctly configured and try again.

DuplicateCallConversionInRequest

Multiple events have the same call and conversion_date_time. Make sure your conversion_date_time is correctly configured and try again.

CustomVariableNotEnabled

Enable the custom variable in your conversion settings and try again.

CustomVariableValueContainsPii

Can’t import events with custom variables containing personally-identifiable information (PII). Remove these variables and try again.

InvalidCustomerForClick

The click from the imported event is associated with a different Google Ads account. Make sure you’re importing to the correct account.

InvalidCustomerForCall

The click from the call is associated with a different Google Ads account. Make sure you’re importing to the correct account. Query conversion_tracking_setting.google_ads_conversion_customer on Customer to identify the correct account.

ConversionNotCompliantWithAttPolicy

The connversion can’t be imported because the conversion source didn’t comply with Apple App Transparency Tracking (ATT) policies or because the customer didn’t consent to tracking.

ClickNotFound

The click can’t be found for the specified identifiers. This may be because it did not come from a Google Ads campaign.

InvalidUserIdentifier

Make sure you hash user provided data using SHA-256 and ensure you are normalizing according to the guidelines.

ExternallyAttributedConversionActionNotPermittedWithUserIdentifier

User provided data can’t be used with external attribution models. Use a different attribution model or omit user identifiers and try again.

UnsupportedUserIdentifier

The provided user identifiers are not supported. Use only hashed email or phone number and try again.

GbraidWbraidBothSet

Can’t use both gbraid and wbraid parameters. Use only 1 and try again.

UnparseableWbraid

Can’t parse event import data. Check if your wbraid parameter was not modified and try again.

UnparseableGbraid

Can’t parse event import data. Check if your gbraid parameter was not modified and try again.

OnePerClickConversionActionNotPermittedWithBraid

Conversion actions that use one-per-click counting can’t be used with gbraid or wbraid parameters.

CustomerDataPolicyProhibitsEnhancedConversions

Enhanced conversions can’t be used for this account because of Google customer data policies. Contact your Google representative.

CustomerNotAcceptedCustomerDataTerms

Make sure you agree to the customer data processing terms in conversion settings and try again. You can check your setting by querying conversion_tracking_setting.accepted_customer_data_terms on Customer.

OrderIdContainsPii

Can’t import events with order IDs containing personally-identifiable information (PII).

CustomerNotEnabledEnhancedConversionsForLeads

Make sure you’ve turned on enhanced conversions for leads in conversion settings and try again. You can check your setting by querying conversion_tracking_setting.enhanced_conversions_for_leads_enabled on Customer.

Implementations

impl ConversionUploadError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionUploadError.

pub fn from_i32(value: i32) -> Option<ConversionUploadError>

Converts an i32 to a ConversionUploadError, or None if value is not a valid variant.

impl ConversionUploadError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.