Enum google_api_proto::google::ads::googleads::v14::errors::image_error_enum::ImageError

#[repr(i32)]
pub enum ImageError {

    Unspecified,
    Unknown,
    InvalidImage,
    StorageError,
    BadRequest,
    UnexpectedSize,
    AnimatedNotAllowed,
    AnimationTooLong,
    ServerError,
    CmykJpegNotAllowed,
    FlashNotAllowed,
    FlashWithoutClicktag,
    FlashErrorAfterFixingClickTag,
    AnimatedVisualEffect,
    FlashError,
    LayoutProblem,
    ProblemReadingImageFile,
    ErrorStoringImage,
    AspectRatioNotAllowed,
    FlashHasNetworkObjects,
    FlashHasNetworkMethods,
    FlashHasUrl,
    FlashHasMouseTracking,
    FlashHasRandomNum,
    FlashSelfTargets,
    FlashBadGeturlTarget,
    FlashVersionNotSupported,
    FlashWithoutHardCodedClickUrl,
    InvalidFlashFile,
    FailedToFixClickTagInFlash,
    FlashAccessesNetworkResources,
    FlashExternalJsCall,
    FlashExternalFsCall,
    FileTooLarge,
    ImageDataTooLarge,
    ImageProcessingError,
    ImageTooSmall,
    InvalidInput,
    ProblemReadingFile,
    ImageConstraintsViolated,
    FormatNotAllowed,

}

Enum describing possible image errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

InvalidImage

The image is not valid.

StorageError

The image could not be stored.

BadRequest

There was a problem with the request.

UnexpectedSize

The image is not of legal dimensions.

AnimatedNotAllowed

Animated image are not permitted.

AnimationTooLong

Animation is too long.

ServerError

There was an error on the server.

CmykJpegNotAllowed

Image cannot be in CMYK color format.

FlashNotAllowed

Flash images are not permitted.

FlashWithoutClicktag

Flash images must support clickTag.

FlashErrorAfterFixingClickTag

A flash error has occurred after fixing the click tag.

AnimatedVisualEffect

Unacceptable visual effects.

FlashError

There was a problem with the flash image.

LayoutProblem

Incorrect image layout.

ProblemReadingImageFile

There was a problem reading the image file.

ErrorStoringImage

There was an error storing the image.

AspectRatioNotAllowed

The aspect ratio of the image is not allowed.

FlashHasNetworkObjects

Flash cannot have network objects.

FlashHasNetworkMethods

Flash cannot have network methods.

FlashHasUrl

Flash cannot have a Url.

FlashHasMouseTracking

Flash cannot use mouse tracking.

FlashHasRandomNum

Flash cannot have a random number.

FlashSelfTargets

Ad click target cannot be ‘_self’.

FlashBadGeturlTarget

GetUrl method should only use ‘_blank’.

FlashVersionNotSupported

Flash version is not supported.

FlashWithoutHardCodedClickUrl

Flash movies need to have hard coded click URL or clickTAG

InvalidFlashFile

Uploaded flash file is corrupted.

FailedToFixClickTagInFlash

Uploaded flash file can be parsed, but the click tag can not be fixed properly.

FlashAccessesNetworkResources

Flash movie accesses network resources

FlashExternalJsCall

Flash movie attempts to call external javascript code

FlashExternalFsCall

Flash movie attempts to call flash system commands

FileTooLarge

Image file is too large.

ImageDataTooLarge

Image data is too large.

ImageProcessingError

Error while processing the image.

ImageTooSmall

Image is too small.

InvalidInput

Input was invalid.

ProblemReadingFile

There was a problem reading the image file.

ImageConstraintsViolated

Image constraints are violated, but details like ASPECT_RATIO_NOT_ALLOWED can’t be provided. This happens when asset spec contains more than one constraint and different criteria of different constraints are violated.

FormatNotAllowed

Image format is not allowed.

Implementations

impl ImageError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ImageError.

pub fn from_i32(value: i32) -> Option<ImageError>

Converts an i32 to a ImageError, or None if value is not a valid variant.

impl ImageError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.