Enum google_api_proto::google::ads::googleads::v14::errors::offline_user_data_job_error_enum::OfflineUserDataJobError

#[repr(i32)]
pub enum OfflineUserDataJobError {

    Unspecified,
    Unknown,
    InvalidUserListId,
    InvalidUserListType,
    NotOnAllowlistForUserId,
    IncompatibleUploadKeyType,
    MissingUserIdentifier,
    InvalidMobileIdFormat,
    TooManyUserIdentifiers,
    NotOnAllowlistForStoreSalesDirect,
    NotOnAllowlistForUnifiedStoreSales,
    InvalidPartnerId,
    InvalidEncoding,
    InvalidCountryCode,
    IncompatibleUserIdentifier,
    FutureTransactionTime,
    InvalidConversionAction,
    MobileIdNotSupported,
    InvalidOperationOrder,
    ConflictingOperation,
    ExternalUpdateIdAlreadyExists,
    JobAlreadyStarted,
    RemoveNotSupported,
    RemoveAllNotSupported,
    InvalidSha256Format,
    CustomKeyDisabled,
    CustomKeyNotPredefined,
    CustomKeyNotSet,
    CustomerNotAcceptedCustomerDataTerms,
    AttributesNotApplicableForCustomerMatchUserList,
    LifetimeValueBucketNotInRange,
    IncompatibleUserIdentifierForAttributes,
    FutureTimeNotAllowed,
    LastPurchaseTimeLessThanAcquisitionTime,
    CustomerIdentifierNotAllowed,
    InvalidItemId,
    FirstPurchaseTimeGreaterThanLastPurchaseTime,
    InvalidLifecycleStage,
    InvalidEventValue,
    EventAttributeAllFieldsAreRequired,

}

Enum describing possible request errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

InvalidUserListId

The user list ID provided for the job is invalid.

InvalidUserListType

Type of the user list is not applicable for the job.

NotOnAllowlistForUserId

Customer is not allowisted for using user ID in upload data.

IncompatibleUploadKeyType

Upload data is not compatible with the upload key type of the associated user list.

MissingUserIdentifier

The user identifier is missing valid data.

InvalidMobileIdFormat

The mobile ID is malformed.

TooManyUserIdentifiers

Maximum number of user identifiers allowed per request is 100,000 and per operation is 20.

NotOnAllowlistForStoreSalesDirect

Customer is not on the allow-list for store sales direct data.

NotOnAllowlistForUnifiedStoreSales

Customer is not on the allow-list for unified store sales data.

InvalidPartnerId

The partner ID in store sales direct metadata is invalid.

InvalidEncoding

The data in user identifier should not be encoded.

InvalidCountryCode

The country code is invalid.

IncompatibleUserIdentifier

Incompatible user identifier when using third_party_user_id for store sales direct first party data or not using third_party_user_id for store sales third party data.

FutureTransactionTime

A transaction time in the future is not allowed.

InvalidConversionAction

The conversion_action specified in transaction_attributes is used to report conversions to a conversion action configured in Google Ads. This error indicates there is no such conversion action in the account.

MobileIdNotSupported

Mobile ID is not supported for store sales direct data.

InvalidOperationOrder

When a remove-all operation is provided, it has to be the first operation of the operation list.

ConflictingOperation

Mixing creation and removal of offline data in the same job is not allowed.

ExternalUpdateIdAlreadyExists

The external update ID already exists.

JobAlreadyStarted

Once the upload job is started, new operations cannot be added.

RemoveNotSupported

Remove operation is not allowed for store sales direct updates.

RemoveAllNotSupported

Remove-all is not supported for certain offline user data job types.

InvalidSha256Format

The SHA256 encoded value is malformed.

CustomKeyDisabled

The custom key specified is not enabled for the unified store sales upload.

CustomKeyNotPredefined

The custom key specified is not predefined through the Google Ads UI.

CustomKeyNotSet

The custom key specified is not set in the upload.

CustomerNotAcceptedCustomerDataTerms

The customer has not accepted the customer data terms in the conversion settings page.

AttributesNotApplicableForCustomerMatchUserList

User attributes cannot be uploaded into a user list.

LifetimeValueBucketNotInRange

Lifetime bucket value must be a number from 0 to 10; 0 is only accepted for remove operations

IncompatibleUserIdentifierForAttributes

Identifiers not supported for Customer Match attributes. User attributes can only be provided with contact info (email, phone, address) user identifiers.

FutureTimeNotAllowed

A time in the future is not allowed.

LastPurchaseTimeLessThanAcquisitionTime

Last purchase date time cannot be less than acquisition date time.

CustomerIdentifierNotAllowed

Only emails are accepted as user identifiers for shopping loyalty match. {– api.dev/not-precedent: The identifier is not limited to ids, but also include other user info eg. phone numbers.}

InvalidItemId

Provided item ID is invalid.

FirstPurchaseTimeGreaterThanLastPurchaseTime

First purchase date time cannot be greater than the last purchase date time.

InvalidLifecycleStage

Provided lifecycle stage is invalid.

InvalidEventValue

The event value of the Customer Match user attribute is invalid.

EventAttributeAllFieldsAreRequired

All the fields are not present in the EventAttribute of the Customer Match.

Implementations

impl OfflineUserDataJobError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of OfflineUserDataJobError.

pub fn from_i32(value: i32) -> Option<OfflineUserDataJobError>

Converts an i32 to a OfflineUserDataJobError, or None if value is not a valid variant.

impl OfflineUserDataJobError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.