Enum google_api_proto::google::ads::googleads::v14::errors::query_error_enum::QueryError

#[repr(i32)]
pub enum QueryError {

    Unspecified = 0,
    Unknown = 1,
    QueryError = 50,
    BadEnumConstant = 18,
    BadEscapeSequence = 7,
    BadFieldName = 12,
    BadLimitValue = 15,
    BadNumber = 5,
    BadOperator = 3,
    BadParameterName = 61,
    BadParameterValue = 62,
    BadResourceTypeInFromClause = 45,
    BadSymbol = 2,
    BadValue = 4,
    DateRangeTooWide = 36,
    DateRangeTooNarrow = 60,
    ExpectedAnd = 30,
    ExpectedBy = 14,
    ExpectedDimensionFieldInSelectClause = 37,
    ExpectedFiltersOnDateRange = 55,
    ExpectedFrom = 44,
    ExpectedList = 41,
    ExpectedReferencedFieldInSelectClause = 16,
    ExpectedSelect = 13,
    ExpectedSingleValue = 42,
    ExpectedValueWithBetweenOperator = 29,
    InvalidDateFormat = 38,
    MisalignedDateForFilter = 64,
    InvalidStringValue = 57,
    InvalidValueWithBetweenOperator = 26,
    InvalidValueWithDuringOperator = 22,
    InvalidValueWithLikeOperator = 56,
    OperatorFieldMismatch = 35,
    ProhibitedEmptyListInCondition = 28,
    ProhibitedEnumConstant = 54,
    ProhibitedFieldCombinationInSelectClause = 31,
    ProhibitedFieldInOrderByClause = 40,
    ProhibitedFieldInSelectClause = 23,
    ProhibitedFieldInWhereClause = 24,
    ProhibitedResourceTypeInFromClause = 43,
    ProhibitedResourceTypeInSelectClause = 48,
    ProhibitedResourceTypeInWhereClause = 58,
    ProhibitedMetricInSelectOrWhereClause = 49,
    ProhibitedSegmentInSelectOrWhereClause = 51,
    ProhibitedSegmentWithMetricInSelectOrWhereClause = 53,
    LimitValueTooLow = 25,
    ProhibitedNewlineInString = 8,
    ProhibitedValueCombinationInList = 10,
    ProhibitedValueCombinationWithBetweenOperator = 21,
    StringNotTerminated = 6,
    TooManySegments = 34,
    UnexpectedEndOfQuery = 9,
    UnexpectedFromClause = 47,
    UnrecognizedField = 32,
    UnexpectedInput = 11,
    RequestedMetricsForManager = 59,
    FilterHasTooManyValues = 63,

}

Enum describing possible query errors.

Variants

Unspecified = 0

Name unspecified.

Unknown = 1

The received error code is not known in this version.

QueryError = 50

Returned if all other query error reasons are not applicable.

BadEnumConstant = 18

A condition used in the query references an invalid enum constant.

BadEscapeSequence = 7

Query contains an invalid escape sequence.

BadFieldName = 12

Field name is invalid.

BadLimitValue = 15

Limit value is invalid (for example, not a number)

BadNumber = 5

Encountered number can not be parsed.

BadOperator = 3

Invalid operator encountered.

BadParameterName = 61

Parameter unknown or not supported.

BadParameterValue = 62

Parameter have invalid value.

BadResourceTypeInFromClause = 45

Invalid resource type was specified in the FROM clause.

BadSymbol = 2

Non-ASCII symbol encountered outside of strings.

BadValue = 4

Value is invalid.

DateRangeTooWide = 36

Date filters fail to restrict date to a range smaller than 31 days. Applicable if the query is segmented by date.

DateRangeTooNarrow = 60

Filters on date/week/month/quarter have a start date after end date.

ExpectedAnd = 30

Expected AND between values with BETWEEN operator.

ExpectedBy = 14

Expecting ORDER BY to have BY.

ExpectedDimensionFieldInSelectClause = 37

There was no dimension field selected.

ExpectedFiltersOnDateRange = 55

Missing filters on date related fields.

ExpectedFrom = 44

Missing FROM clause.

ExpectedList = 41

The operator used in the conditions requires the value to be a list.

ExpectedReferencedFieldInSelectClause = 16

Fields used in WHERE or ORDER BY clauses are missing from the SELECT clause.

ExpectedSelect = 13

SELECT is missing at the beginning of query.

ExpectedSingleValue = 42

A list was passed as a value to a condition whose operator expects a single value.

ExpectedValueWithBetweenOperator = 29

Missing one or both values with BETWEEN operator.

InvalidDateFormat = 38

Invalid date format. Expected โYYYY-MM-DDโ.

MisalignedDateForFilter = 64

Misaligned date value for the filter. The date should be the start of a week/month/quarter if the filtered field is segments.week/segments.month/segments.quarter.

InvalidStringValue = 57

Value passed was not a string when it should have been. For example, it was a number or unquoted literal.

InvalidValueWithBetweenOperator = 26

A String value passed to the BETWEEN operator does not parse as a date.

InvalidValueWithDuringOperator = 22

The value passed to the DURING operator is not a Date range literal

InvalidValueWithLikeOperator = 56

An invalid value was passed to the LIKE operator.

OperatorFieldMismatch = 35

An operator was provided that is inapplicable to the field being filtered.

ProhibitedEmptyListInCondition = 28

A Condition was found with an empty list.

ProhibitedEnumConstant = 54

A condition used in the query references an unsupported enum constant.

ProhibitedFieldCombinationInSelectClause = 31

Fields that are not allowed to be selected together were included in the SELECT clause.

ProhibitedFieldInOrderByClause = 40

A field that is not orderable was included in the ORDER BY clause.

ProhibitedFieldInSelectClause = 23

A field that is not selectable was included in the SELECT clause.

ProhibitedFieldInWhereClause = 24

A field that is not filterable was included in the WHERE clause.

ProhibitedResourceTypeInFromClause = 43

Resource type specified in the FROM clause is not supported by this service.

ProhibitedResourceTypeInSelectClause = 48

A field that comes from an incompatible resource was included in the SELECT clause.

ProhibitedResourceTypeInWhereClause = 58

A field that comes from an incompatible resource was included in the WHERE clause.

ProhibitedMetricInSelectOrWhereClause = 49

A metric incompatible with the main resource or other selected segmenting resources was included in the SELECT or WHERE clause.

ProhibitedSegmentInSelectOrWhereClause = 51

A segment incompatible with the main resource or other selected segmenting resources was included in the SELECT or WHERE clause.

ProhibitedSegmentWithMetricInSelectOrWhereClause = 53

A segment in the SELECT clause is incompatible with a metric in the SELECT or WHERE clause.

LimitValueTooLow = 25

The value passed to the limit clause is too low.

ProhibitedNewlineInString = 8

Query has a string containing a newline character.

ProhibitedValueCombinationInList = 10

List contains values of different types.

ProhibitedValueCombinationWithBetweenOperator = 21

The values passed to the BETWEEN operator are not of the same type.

StringNotTerminated = 6

Query contains unterminated string.

TooManySegments = 34

Too many segments are specified in SELECT clause.

UnexpectedEndOfQuery = 9

Query is incomplete and cannot be parsed.

UnexpectedFromClause = 47

FROM clause cannot be specified in this query.

UnrecognizedField = 32

Query contains one or more unrecognized fields.

UnexpectedInput = 11

Query has an unexpected extra part.

RequestedMetricsForManager = 59

Metrics cannot be requested for a manager account. To retrieve metrics, issue separate requests against each client account under the manager account.

FilterHasTooManyValues = 63

The number of values (right-hand-side operands) in a filter exceeds the limit.

Implementations

impl QueryError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of QueryError.

pub fn from_i32(value: i32) -> Option<QueryError>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a QueryError, or None if value is not a valid variant.

impl QueryError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.