Module google_api_proto::google::ads::googleads::v14::resources::recommendation

Nested message and enum types in Recommendation.

Modules

campaign_budget_recommendation

Nested message and enum types in CampaignBudgetRecommendation.

keyword_recommendation

Nested message and enum types in KeywordRecommendation.

target_cpa_opt_in_recommendation

Nested message and enum types in TargetCpaOptInRecommendation.

Structs

CallAssetRecommendation

The call asset recommendation.

CalloutAssetRecommendation

The callout asset recommendation.

CampaignBudget

A campaign budget shared amongst various budget recommendation types.

CampaignBudgetRecommendation

The budget recommendation for budget constrained campaigns.

DisplayExpansionOptInRecommendation

The Display Expansion opt-in recommendation.

DynamicImageExtensionOptInRecommendation

Recommendation to enable dynamic image extensions on the account, allowing Google to find the best images from ad landing pages and complement text ads.

EnhancedCpcOptInRecommendation

The Enhanced Cost-Per-Click Opt-In recommendation.

ForecastingSetTargetRoasRecommendation

The forecasting set target ROAS recommendation.

ImprovePerformanceMaxAdStrengthRecommendation

Recommendation to improve the asset group strength of a Performance Max campaign to an “Excellent” rating.

KeywordMatchTypeRecommendation

The keyword match type recommendation.

KeywordRecommendation

The keyword recommendation.

LowerTargetRoasRecommendation

Recommendation to lower Target ROAS.

MaximizeClicksOptInRecommendation

The Maximize Clicks opt-in recommendation.

MaximizeConversionsOptInRecommendation

The Maximize Conversions Opt-In recommendation.

MerchantInfo

The Merchant Center account details.

MigrateDynamicSearchAdsCampaignToPerformanceMaxRecommendation

The Dynamic Search Ads to Performance Max migration recommendation.

MoveUnusedBudgetRecommendation

The move unused budget recommendation.

OptimizeAdRotationRecommendation

The Optimize Ad Rotation recommendation.

PerformanceMaxOptInRecommendation

The Performance Max Opt In recommendation.

RaiseTargetCpaBidTooLowRecommendation

The raise target CPA bid too low recommendation.

RaiseTargetCpaRecommendation

Recommendation to raise Target CPA.

RecommendationImpact

The impact of making the change as described in the recommendation. Some types of recommendations may not have impact information.

RecommendationMetrics

Weekly account performance metrics. For some recommendation types, these are averaged over the past 90-day period and hence can be fractional.

ResponsiveSearchAdAssetRecommendation

The add responsive search ad asset recommendation.

ResponsiveSearchAdImproveAdStrengthRecommendation

The responsive search ad improve ad strength recommendation.

ResponsiveSearchAdRecommendation

The add responsive search ad recommendation.

SearchPartnersOptInRecommendation

The Search Partners Opt-In recommendation.

ShoppingAddProductsToCampaignRecommendation

The shopping recommendation to add products to a Shopping Campaign Inventory.

ShoppingFixDisapprovedProductsRecommendation

The shopping recommendation to fix disapproved products in a Shopping Campaign Inventory.

ShoppingMerchantCenterAccountSuspensionRecommendation

The shopping recommendation to fix Merchant Center account suspension issues.

ShoppingMigrateRegularShoppingCampaignOffersToPerformanceMaxRecommendation

The shopping recommendation to migrate Regular Shopping Campaign targeted offers to Performance Max campaigns.

ShoppingOfferAttributeRecommendation

The shopping recommendation to add an attribute to offers that are demoted because it is missing.

ShoppingTargetAllOffersRecommendation

The shopping recommendation to create a catch-all campaign that targets all offers.

SitelinkAssetRecommendation

The sitelink asset recommendation.

TargetAdjustmentInfo

Information of a target adjustment recommendation.

TargetCpaOptInRecommendation

The Target CPA opt-in recommendation.

TargetRoasOptInRecommendation

The Target ROAS opt-in recommendation.

TextAdRecommendation

The text ad recommendation.

UpgradeLocalCampaignToPerformanceMaxRecommendation

The Upgrade Local campaign to Performance Max campaign recommendation.

UpgradeSmartShoppingCampaignToPerformanceMaxRecommendation

The upgrade a Smart Shopping campaign to a Performance Max campaign recommendation.

UseBroadMatchKeywordRecommendation

The use broad match keyword recommendation.

Enums

Recommendation

The details of recommendation.