Struct google_api_proto::google::ads::googleads::v14::resources::AdGroup

pub struct AdGroup {

    pub resource_name: String,
    pub id: Option<i64>,
    pub name: Option<String>,
    pub status: i32,
    pub type: i32,
    pub ad_rotation_mode: i32,
    pub base_ad_group: Option<String>,
    pub tracking_url_template: Option<String>,
    pub url_custom_parameters: Vec<CustomParameter>,
    pub campaign: Option<String>,
    pub cpc_bid_micros: Option<i64>,
    pub effective_cpc_bid_micros: Option<i64>,
    pub cpm_bid_micros: Option<i64>,
    pub target_cpa_micros: Option<i64>,
    pub cpv_bid_micros: Option<i64>,
    pub target_cpm_micros: Option<i64>,
    pub target_roas: Option<f64>,
    pub percent_cpc_bid_micros: Option<i64>,
    pub optimized_targeting_enabled: bool,
    pub display_custom_bid_dimension: i32,
    pub final_url_suffix: Option<String>,
    pub targeting_setting: Option<TargetingSetting>,
    pub audience_setting: Option<AudienceSetting>,
    pub effective_target_cpa_micros: Option<i64>,
    pub effective_target_cpa_source: i32,
    pub effective_target_roas: Option<f64>,
    pub effective_target_roas_source: i32,
    pub labels: Vec<String>,
    pub excluded_parent_asset_field_types: Vec<i32>,
    pub excluded_parent_asset_set_types: Vec<i32>,

}

An ad group.

Fields
resource_name: String

Immutable. The resource name of the ad group. Ad group resource names have the form:

customers/{customer_id}/adGroups/{ad_group_id}

id: Option<i64>

Output only. The ID of the ad group.

name: Option<String>

The name of the ad group.

This field is required and should not be empty when creating new ad groups.

It must contain fewer than 255 UTF-8 full-width characters.

It must not contain any null (code point 0x0), NL line feed (code point 0xA) or carriage return (code point 0xD) characters.

status: i32

The status of the ad group.

type: i32

Immutable. The type of the ad group.

ad_rotation_mode: i32

The ad rotation mode of the ad group.

base_ad_group: Option<String>

Output only. For draft or experiment ad groups, this field is the resource name of the base ad group from which this ad group was created. If a draft or experiment ad group does not have a base ad group, then this field is null.

For base ad groups, this field equals the ad group resource name.

This field is read-only.

tracking_url_template: Option<String>

The URL template for constructing a tracking URL.

url_custom_parameters: Vec<CustomParameter>

The list of mappings used to substitute custom parameter tags in a tracking_url_template, final_urls, or mobile_final_urls.

campaign: Option<String>

Immutable. The campaign to which the ad group belongs.

cpc_bid_micros: Option<i64>

The maximum CPC (cost-per-click) bid.

effective_cpc_bid_micros: Option<i64>

Output only. Value will be same as that of the CPC (cost-per-click) bid value when the bidding strategy is one of manual cpc, enhanced cpc, page one promoted or target outrank share, otherwise the value will be null.

cpm_bid_micros: Option<i64>

The maximum CPM (cost-per-thousand viewable impressions) bid.

target_cpa_micros: Option<i64>

The target CPA (cost-per-acquisition). If the ad group’s campaign bidding strategy is TargetCpa or MaximizeConversions (with its target_cpa field set), then this field overrides the target CPA specified in the campaign’s bidding strategy. Otherwise, this value is ignored.

cpv_bid_micros: Option<i64>

The CPV (cost-per-view) bid.

target_cpm_micros: Option<i64>

Average amount in micros that the advertiser is willing to pay for every thousand times the ad is shown.

target_roas: Option<f64>

The target ROAS (return-on-ad-spend) override. If the ad group’s campaign bidding strategy is TargetRoas or MaximizeConversionValue (with its target_roas field set), then this field overrides the target ROAS specified in the campaign’s bidding strategy. Otherwise, this value is ignored.

percent_cpc_bid_micros: Option<i64>

The percent cpc bid amount, expressed as a fraction of the advertised price for some good or service. The valid range for the fraction is [0,1) and the value stored here is 1,000,000 * [fraction].

optimized_targeting_enabled: bool

True if optimized targeting is enabled. Optimized Targeting is the replacement for Audience Expansion.

display_custom_bid_dimension: i32

Allows advertisers to specify a targeting dimension on which to place absolute bids. This is only applicable for campaigns that target only the display network and not search.

final_url_suffix: Option<String>

URL template for appending params to Final URL.

targeting_setting: Option<TargetingSetting>

Setting for targeting related features.

audience_setting: Option<AudienceSetting>

Immutable. Setting for audience related features.

effective_target_cpa_micros: Option<i64>

Output only. The effective target CPA (cost-per-acquisition). This field is read-only.

effective_target_cpa_source: i32

Output only. Source of the effective target CPA. This field is read-only.

effective_target_roas: Option<f64>

Output only. The effective target ROAS (return-on-ad-spend). This field is read-only.

effective_target_roas_source: i32

Output only. Source of the effective target ROAS. This field is read-only.

labels: Vec<String>

Output only. The resource names of labels attached to this ad group.

excluded_parent_asset_field_types: Vec<i32>

The asset field types that should be excluded from this ad group. Asset links with these field types will not be inherited by this ad group from the upper levels.

excluded_parent_asset_set_types: Vec<i32>

The asset set types that should be excluded from this ad group. Asset set links with these types will not be inherited by this ad group from the upper levels. Location group types (GMB_DYNAMIC_LOCATION_GROUP, CHAIN_DYNAMIC_LOCATION_GROUP, and STATIC_LOCATION_GROUP) are child types of LOCATION_SYNC. Therefore, if LOCATION_SYNC is set for this field, all location group asset sets are not allowed to be linked to this ad group, and all Location Extension (LE) and Affiliate Location Extensions (ALE) will not be served under this ad group. Only LOCATION_SYNC is currently supported.

Implementations

impl AdGroup

pub fn status(&self) -> AdGroupStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: AdGroupStatus)

Sets status to the provided enum value.

pub fn type(&self) -> AdGroupType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: AdGroupType)

Sets type to the provided enum value.

pub fn ad_rotation_mode(&self) -> AdGroupAdRotationMode

Returns the enum value of ad_rotation_mode, or the default if the field is set to an invalid enum value.

pub fn set_ad_rotation_mode(&mut self, value: AdGroupAdRotationMode)

Sets ad_rotation_mode to the provided enum value.

pub fn display_custom_bid_dimension(&self) -> TargetingDimension

Returns the enum value of display_custom_bid_dimension, or the default if the field is set to an invalid enum value.

pub fn set_display_custom_bid_dimension(&mut self, value: TargetingDimension)

Sets display_custom_bid_dimension to the provided enum value.

pub fn effective_target_cpa_source(&self) -> BiddingSource

Returns the enum value of effective_target_cpa_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_target_cpa_source(&mut self, value: BiddingSource)

Sets effective_target_cpa_source to the provided enum value.

pub fn effective_target_roas_source(&self) -> BiddingSource

Returns the enum value of effective_target_roas_source, or the default if the field is set to an invalid enum value.

pub fn set_effective_target_roas_source(&mut self, value: BiddingSource)

Sets effective_target_roas_source to the provided enum value.

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn base_ad_group(&self) -> &str

Returns the value of base_ad_group, or the default value if base_ad_group is unset.

pub fn tracking_url_template(&self) -> &str

Returns the value of tracking_url_template, or the default value if tracking_url_template is unset.

pub fn campaign(&self) -> &str

Returns the value of campaign, or the default value if campaign is unset.

pub fn cpc_bid_micros(&self) -> i64

Returns the value of cpc_bid_micros, or the default value if cpc_bid_micros is unset.

pub fn cpm_bid_micros(&self) -> i64

Returns the value of cpm_bid_micros, or the default value if cpm_bid_micros is unset.

pub fn target_cpa_micros(&self) -> i64

Returns the value of target_cpa_micros, or the default value if target_cpa_micros is unset.

pub fn cpv_bid_micros(&self) -> i64

Returns the value of cpv_bid_micros, or the default value if cpv_bid_micros is unset.

pub fn target_cpm_micros(&self) -> i64

Returns the value of target_cpm_micros, or the default value if target_cpm_micros is unset.

pub fn target_roas(&self) -> f64

Returns the value of target_roas, or the default value if target_roas is unset.

pub fn percent_cpc_bid_micros(&self) -> i64

Returns the value of percent_cpc_bid_micros, or the default value if percent_cpc_bid_micros is unset.

pub fn final_url_suffix(&self) -> &str

Returns the value of final_url_suffix, or the default value if final_url_suffix is unset.

pub fn effective_target_cpa_micros(&self) -> i64

Returns the value of effective_target_cpa_micros, or the default value if effective_target_cpa_micros is unset.

pub fn effective_target_roas(&self) -> f64

Returns the value of effective_target_roas, or the default value if effective_target_roas is unset.

pub fn excluded_parent_asset_field_types( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<AssetFieldType>>

Returns an iterator which yields the valid enum values contained in excluded_parent_asset_field_types.

pub fn push_excluded_parent_asset_field_types(&mut self, value: AssetFieldType)

Appends the provided enum value to excluded_parent_asset_field_types.

pub fn effective_cpc_bid_micros(&self) -> i64

Returns the value of effective_cpc_bid_micros, or the default value if effective_cpc_bid_micros is unset.

pub fn excluded_parent_asset_set_types( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<AssetSetType>>

Returns an iterator which yields the valid enum values contained in excluded_parent_asset_set_types.

pub fn push_excluded_parent_asset_set_types(&mut self, value: AssetSetType)

Appends the provided enum value to excluded_parent_asset_set_types.