Enum google_api_proto::google::ads::googleads::v15::enums::placeholder_type_enum::PlaceholderType

#[repr(i32)]
pub enum PlaceholderType {

    Unspecified = 0,
    Unknown = 1,
    Sitelink = 2,
    Call = 3,
    App = 4,
    Location = 5,
    AffiliateLocation = 6,
    Callout = 7,
    StructuredSnippet = 8,
    Message = 9,
    Price = 10,
    Promotion = 11,
    AdCustomizer = 12,
    DynamicEducation = 13,
    DynamicFlight = 14,
    DynamicCustom = 15,
    DynamicHotel = 16,
    DynamicRealEstate = 17,
    DynamicTravel = 18,
    DynamicLocal = 19,
    DynamicJob = 20,
    Image = 21,

}

Possible placeholder types for a feed mapping.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

Sitelink = 2

Lets you show links in your ad to pages from your website, including the main landing page.

Call = 3

Lets you attach a phone number to an ad, allowing customers to call directly from the ad.

App = 4

Lets you provide users with a link that points to a mobile app in addition to a website.

Location = 5

Lets you show locations of businesses from your Business Profile in your ad. This helps people find your locations by showing your ads with your address, a map to your location, or the distance to your business. This extension type is useful to draw customers to your brick-and-mortar location.

AffiliateLocation = 6

If you sell your product through retail chains, affiliate location extensions let you show nearby stores that carry your products.

Callout = 7

Lets you include additional text with your search ads that provide detailed information about your business, including products and services you offer. Callouts appear in ads at the top and bottom of Google search results.

StructuredSnippet = 8

Lets you add more info to your ad, specific to some predefined categories such as types, brands, styles, etc. A minimum of 3 text (SNIPPETS) values are required.

Message = 9

Allows users to see your ad, click an icon, and contact you directly by text message. With one tap on your ad, people can contact you to book an appointment, get a quote, ask for information, or request a service.

Price = 10

Lets you display prices for a list of items along with your ads. A price feed is composed of three to eight price table rows.

Promotion = 11

Lets you highlight sales and other promotions that let users see how they can save by buying now.

AdCustomizer = 12

Lets you dynamically inject custom data into the title and description of your ads.

DynamicEducation = 13

Indicates that this feed is for education dynamic remarketing.

DynamicFlight = 14

Indicates that this feed is for flight dynamic remarketing.

DynamicCustom = 15

Indicates that this feed is for a custom dynamic remarketing type. Use this only if the other business types don’t apply to your products or services.

DynamicHotel = 16

Indicates that this feed is for hotels and rentals dynamic remarketing.

DynamicRealEstate = 17

Indicates that this feed is for real estate dynamic remarketing.

DynamicTravel = 18

Indicates that this feed is for travel dynamic remarketing.

DynamicLocal = 19

Indicates that this feed is for local deals dynamic remarketing.

DynamicJob = 20

Indicates that this feed is for job dynamic remarketing.

Image = 21

Lets you attach an image to an ad.

Implementations

impl PlaceholderType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of PlaceholderType.

pub fn from_i32(value: i32) -> Option<PlaceholderType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a PlaceholderType, or None if value is not a valid variant.

impl PlaceholderType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.