Enum google_api_proto::google::ads::googleads::v15::enums::promotion_extension_occasion_enum::PromotionExtensionOccasion

#[repr(i32)]
pub enum PromotionExtensionOccasion {

    Unspecified = 0,
    Unknown = 1,
    NewYears = 2,
    ChineseNewYear = 3,
    ValentinesDay = 4,
    Easter = 5,
    MothersDay = 6,
    FathersDay = 7,
    LaborDay = 8,
    BackToSchool = 9,
    Halloween = 10,
    BlackFriday = 11,
    CyberMonday = 12,
    Christmas = 13,
    BoxingDay = 14,
    IndependenceDay = 15,
    NationalDay = 16,
    EndOfSeason = 17,
    WinterSale = 18,
    SummerSale = 19,
    FallSale = 20,
    SpringSale = 21,
    Ramadan = 22,
    EidAlFitr = 23,
    EidAlAdha = 24,
    SinglesDay = 25,
    WomensDay = 26,
    Holi = 27,
    ParentsDay = 28,
    StNicholasDay = 29,
    Carnival = 30,
    Epiphany = 31,
    RoshHashanah = 32,
    Passover = 33,
    Hanukkah = 34,
    Diwali = 35,
    Navratri = 36,
    Songkran = 37,
    YearEndGift = 38,

}

A promotion extension occasion.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

NewYears = 2

New Yearโs.

ChineseNewYear = 3

Chinese New Year.

ValentinesDay = 4

Valentineโs Day.

Easter = 5

Easter.

MothersDay = 6

Motherโs Day.

FathersDay = 7

Fatherโs Day.

LaborDay = 8

Labor Day.

BackToSchool = 9

Back To School.

Halloween = 10

Halloween.

BlackFriday = 11

Black Friday.

CyberMonday = 12

Cyber Monday.

Christmas = 13

Christmas.

BoxingDay = 14

Boxing Day.

IndependenceDay = 15

Independence Day in any country.

NationalDay = 16

National Day in any country.

EndOfSeason = 17

End of any season.

WinterSale = 18

Winter Sale.

SummerSale = 19

Summer sale.

FallSale = 20

Fall Sale.

SpringSale = 21

Spring Sale.

Ramadan = 22

Ramadan.

EidAlFitr = 23

Eid al-Fitr.

EidAlAdha = 24

Eid al-Adha.

SinglesDay = 25

Singles Day.

WomensDay = 26

Womenโs Day.

Holi = 27

Holi.

ParentsDay = 28

Parentโs Day.

StNicholasDay = 29

St. Nicholas Day.

Carnival = 30

Carnival.

Epiphany = 31

Epiphany, also known as Three Kingsโ Day.

RoshHashanah = 32

Rosh Hashanah.

Passover = 33

Passover.

Hanukkah = 34

Hanukkah.

Diwali = 35

Diwali.

Navratri = 36

Navratri.

Songkran = 37

Available in Thai: Songkran.

YearEndGift = 38

Available in Japanese: Year-end Gift.

Implementations

impl PromotionExtensionOccasion

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of PromotionExtensionOccasion.

pub fn from_i32(value: i32) -> Option<PromotionExtensionOccasion>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a PromotionExtensionOccasion, or None if value is not a valid variant.

impl PromotionExtensionOccasion

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.