Enum google_api_proto::google::ads::googleads::v15::errors::ad_error_enum::AdError

#[repr(i32)]
pub enum AdError {

    Unspecified = 0,
    Unknown = 1,
    AdCustomizersNotSupportedForAdType = 2,
    ApproximatelyTooLong = 3,
    ApproximatelyTooShort = 4,
    BadSnippet = 5,
    CannotModifyAd = 6,
    CannotSetBusinessNameIfUrlSet = 7,
    CannotSetField = 8,
    CannotSetFieldWithOriginAdIdSet = 9,
    CannotSetFieldWithAdIdSetForSharing = 10,
    CannotSetAllowFlexibleColorFalse = 11,
    CannotSetColorControlWhenNativeFormatSetting = 12,
    CannotSetUrl = 13,
    CannotSetWithoutFinalUrls = 14,
    CannotSetWithFinalUrls = 15,
    CannotSetWithUrlData = 17,
    CannotUseAdSubclassForOperator = 18,
    CustomerNotApprovedMobileads = 19,
    CustomerNotApprovedThirdpartyAds = 20,
    CustomerNotApprovedThirdpartyRedirectAds = 21,
    CustomerNotEligible = 22,
    CustomerNotEligibleForUpdatingBeaconUrl = 23,
    DimensionAlreadyInUnion = 24,
    DimensionMustBeSet = 25,
    DimensionNotInUnion = 26,
    DisplayUrlCannotBeSpecified = 27,
    DomesticPhoneNumberFormat = 28,
    EmergencyPhoneNumber = 29,
    EmptyField = 30,
    FeedAttributeMustHaveMappingForTypeId = 31,
    FeedAttributeMappingTypeMismatch = 32,
    IllegalAdCustomizerTagUse = 33,
    IllegalTagUse = 34,
    InconsistentDimensions = 35,
    InconsistentStatusInTemplateUnion = 36,
    IncorrectLength = 37,
    IneligibleForUpgrade = 38,
    InvalidAdAddressCampaignTarget = 39,
    InvalidAdType = 40,
    InvalidAttributesForMobileImage = 41,
    InvalidAttributesForMobileText = 42,
    InvalidCallToActionText = 43,
    InvalidCharacterForUrl = 44,
    InvalidCountryCode = 45,
    InvalidExpandedDynamicSearchAdTag = 47,
    InvalidInput = 48,
    InvalidMarkupLanguage = 49,
    InvalidMobileCarrier = 50,
    InvalidMobileCarrierTarget = 51,
    InvalidNumberOfElements = 52,
    InvalidPhoneNumberFormat = 53,
    InvalidRichMediaCertifiedVendorFormatId = 54,
    InvalidTemplateData = 55,
    InvalidTemplateElementFieldType = 56,
    InvalidTemplateId = 57,
    LineTooWide = 58,
    MissingAdCustomizerMapping = 59,
    MissingAddressComponent = 60,
    MissingAdvertisementName = 61,
    MissingBusinessName = 62,
    MissingDescription1 = 63,
    MissingDescription2 = 64,
    MissingDestinationUrlTag = 65,
    MissingLandingPageUrlTag = 66,
    MissingDimension = 67,
    MissingDisplayUrl = 68,
    MissingHeadline = 69,
    MissingHeight = 70,
    MissingImage = 71,
    MissingMarketingImageOrProductVideos = 72,
    MissingMarkupLanguages = 73,
    MissingMobileCarrier = 74,
    MissingPhone = 75,
    MissingRequiredTemplateFields = 76,
    MissingTemplateFieldValue = 77,
    MissingText = 78,
    MissingVisibleUrl = 79,
    MissingWidth = 80,
    MultipleDistinctFeedsUnsupported = 81,
    MustUseTempAdUnionIdOnAdd = 82,
    TooLong = 83,
    TooShort = 84,
    UnionDimensionsCannotChange = 85,
    UnknownAddressComponent = 86,
    UnknownFieldName = 87,
    UnknownUniqueName = 88,
    UnsupportedDimensions = 89,
    UrlInvalidScheme = 90,
    UrlInvalidTopLevelDomain = 91,
    UrlMalformed = 92,
    UrlNoHost = 93,
    UrlNotEquivalent = 94,
    UrlHostNameTooLong = 95,
    UrlNoScheme = 96,
    UrlNoTopLevelDomain = 97,
    UrlPathNotAllowed = 98,
    UrlPortNotAllowed = 99,
    UrlQueryNotAllowed = 100,
    UrlSchemeBeforeExpandedDynamicSearchAdTag = 102,
    UserDoesNotHaveAccessToTemplate = 103,
    InconsistentExpandableSettings = 104,
    InvalidFormat = 105,
    InvalidFieldText = 106,
    ElementNotPresent = 107,
    ImageError = 108,
    ValueNotInRange = 109,
    FieldNotPresent = 110,
    AddressNotComplete = 111,
    AddressInvalid = 112,
    VideoRetrievalError = 113,
    AudioError = 114,
    InvalidYoutubeDisplayUrl = 115,
    TooManyProductImages = 116,
    TooManyProductVideos = 117,
    IncompatibleAdTypeAndDevicePreference = 118,
    CalltrackingNotSupportedForCountry = 119,
    CarrierSpecificShortNumberNotAllowed = 120,
    DisallowedNumberType = 121,
    PhoneNumberNotSupportedForCountry = 122,
    PhoneNumberNotSupportedWithCalltrackingForCountry = 123,
    PremiumRateNumberNotAllowed = 124,
    VanityPhoneNumberNotAllowed = 125,
    InvalidCallConversionTypeId = 126,
    CannotDisableCallConversionAndSetConversionTypeId = 127,
    CannotSetPath2WithoutPath1 = 128,
    MissingDynamicSearchAdsSettingDomainName = 129,
    IncompatibleWithRestrictionType = 130,
    CustomerConsentForCallRecordingRequired = 131,
    MissingImageOrMediaBundle = 132,
    ProductTypeNotSupportedInThisCampaign = 133,
    PlaceholderCannotHaveEmptyDefaultValue = 134,
    PlaceholderCountdownFunctionCannotHaveDefaultValue = 135,
    PlaceholderDefaultValueMissing = 136,
    UnexpectedPlaceholderDefaultValue = 137,
    AdCustomizersMayNotBeAdjacent = 138,
    UpdatingAdWithNoEnabledAssociation = 139,
    CallAdVerificationUrlFinalUrlDoesNotHaveSameDomain = 140,
    CallAdFinalUrlAndVerificationUrlCannotBothBeEmpty = 154,
    TooManyAdCustomizers = 141,
    InvalidAdCustomizerFormat = 142,
    NestedAdCustomizerSyntax = 143,
    UnsupportedAdCustomizerSyntax = 144,
    UnpairedBraceInAdCustomizerTag = 145,
    MoreThanOneCountdownTagTypeExists = 146,
    DateTimeInCountdownTagIsInvalid = 147,
    DateTimeInCountdownTagIsPast = 148,
    UnrecognizedAdCustomizerTagFound = 149,
    CustomizerTypeForbiddenForField = 150,
    InvalidCustomizerAttributeName = 151,
    StoreMismatch = 152,
    MissingRequiredImageAspectRatio = 153,
    MismatchedAspectRatios = 155,
    DuplicateImageAcrossCarouselCards = 156,

}

Enum describing possible ad errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

AdCustomizersNotSupportedForAdType = 2

Ad customizers are not supported for ad type.

ApproximatelyTooLong = 3

Estimating character sizes the string is too long.

ApproximatelyTooShort = 4

Estimating character sizes the string is too short.

BadSnippet = 5

There is a problem with the snippet.

CannotModifyAd = 6

Cannot modify an ad.

CannotSetBusinessNameIfUrlSet = 7

business name and url cannot be set at the same time

CannotSetField = 8

The specified field is incompatible with this ad’s type or settings.

CannotSetFieldWithOriginAdIdSet = 9

Cannot set field when originAdId is set.

CannotSetFieldWithAdIdSetForSharing = 10

Cannot set field when an existing ad id is set for sharing.

CannotSetAllowFlexibleColorFalse = 11

Cannot set allowFlexibleColor false if no color is provided by user.

CannotSetColorControlWhenNativeFormatSetting = 12

When user select native, no color control is allowed because we will always respect publisher color for native format serving.

CannotSetUrl = 13

Cannot specify a url for the ad type

CannotSetWithoutFinalUrls = 14

Cannot specify a tracking or mobile url without also setting final urls

CannotSetWithFinalUrls = 15

Cannot specify a legacy url and a final url simultaneously

CannotSetWithUrlData = 17

Cannot specify a urls in UrlData and in template fields simultaneously.

CannotUseAdSubclassForOperator = 18

This operator cannot be used with a subclass of Ad.

CustomerNotApprovedMobileads = 19

Customer is not approved for mobile ads.

CustomerNotApprovedThirdpartyAds = 20

Customer is not approved for 3PAS richmedia ads.

CustomerNotApprovedThirdpartyRedirectAds = 21

Customer is not approved for 3PAS redirect richmedia (Ad Exchange) ads.

CustomerNotEligible = 22

Not an eligible customer

CustomerNotEligibleForUpdatingBeaconUrl = 23

Customer is not eligible for updating beacon url

DimensionAlreadyInUnion = 24

There already exists an ad with the same dimensions in the union.

DimensionMustBeSet = 25

Ad’s dimension must be set before setting union dimension.

DimensionNotInUnion = 26

Ad’s dimension must be included in the union dimensions.

DisplayUrlCannotBeSpecified = 27

Display Url cannot be specified (applies to Ad Exchange Ads)

DomesticPhoneNumberFormat = 28

Telephone number contains invalid characters or invalid format. Re-enter your number using digits (0-9), dashes (-), and parentheses only.

EmergencyPhoneNumber = 29

Emergency telephone numbers are not allowed. Enter a valid domestic phone number to connect customers to your business.

EmptyField = 30

A required field was not specified or is an empty string.

FeedAttributeMustHaveMappingForTypeId = 31

A feed attribute referenced in an ad customizer tag is not in the ad customizer mapping for the feed.

FeedAttributeMappingTypeMismatch = 32

The ad customizer field mapping for the feed attribute does not match the expected field type.

IllegalAdCustomizerTagUse = 33

The use of ad customizer tags in the ad text is disallowed. Details in trigger.

IllegalTagUse = 34

Tags of the form {PH_x}, where x is a number, are disallowed in ad text.

InconsistentDimensions = 35

The dimensions of the ad are specified or derived in multiple ways and are not consistent.

InconsistentStatusInTemplateUnion = 36

The status cannot differ among template ads of the same union.

IncorrectLength = 37

The length of the string is not valid.

IneligibleForUpgrade = 38

The ad is ineligible for upgrade.

InvalidAdAddressCampaignTarget = 39

User cannot create mobile ad for countries targeted in specified campaign.

InvalidAdType = 40

Invalid Ad type. A specific type of Ad is required.

InvalidAttributesForMobileImage = 41

Headline, description or phone cannot be present when creating mobile image ad.

InvalidAttributesForMobileText = 42

Image cannot be present when creating mobile text ad.

InvalidCallToActionText = 43

Invalid call to action text.

InvalidCharacterForUrl = 44

Invalid character in URL.

InvalidCountryCode = 45

Creative’s country code is not valid.

InvalidExpandedDynamicSearchAdTag = 47

Invalid use of Expanded Dynamic Search Ads tags ({lpurl} etc.)

InvalidInput = 48

An input error whose real reason was not properly mapped (should not happen).

InvalidMarkupLanguage = 49

An invalid markup language was entered.

InvalidMobileCarrier = 50

An invalid mobile carrier was entered.

InvalidMobileCarrierTarget = 51

Specified mobile carriers target a country not targeted by the campaign.

InvalidNumberOfElements = 52

Wrong number of elements for given element type

InvalidPhoneNumberFormat = 53

The format of the telephone number is incorrect. Re-enter the number using the correct format.

InvalidRichMediaCertifiedVendorFormatId = 54

The certified vendor format id is incorrect.

InvalidTemplateData = 55

The template ad data contains validation errors.

InvalidTemplateElementFieldType = 56

The template field doesn’t have have the correct type.

InvalidTemplateId = 57

Invalid template id.

LineTooWide = 58

After substituting replacement strings, the line is too wide.

MissingAdCustomizerMapping = 59

The feed referenced must have ad customizer mapping to be used in a customizer tag.

MissingAddressComponent = 60

Missing address component in template element address field.

MissingAdvertisementName = 61

An ad name must be entered.

MissingBusinessName = 62

Business name must be entered.

MissingDescription1 = 63

Description (line 2) must be entered.

MissingDescription2 = 64

Description (line 3) must be entered.

MissingDestinationUrlTag = 65

The destination url must contain at least one tag (for example, {lpurl})

MissingLandingPageUrlTag = 66

The tracking url template of ExpandedDynamicSearchAd must contain at least one tag. (for example, {lpurl})

MissingDimension = 67

A valid dimension must be specified for this ad.

MissingDisplayUrl = 68

A display URL must be entered.

MissingHeadline = 69

Headline must be entered.

MissingHeight = 70

A height must be entered.

MissingImage = 71

An image must be entered.

MissingMarketingImageOrProductVideos = 72

Marketing image or product videos are required.

MissingMarkupLanguages = 73

The markup language in which your site is written must be entered.

MissingMobileCarrier = 74

A mobile carrier must be entered.

MissingPhone = 75

Phone number must be entered.

MissingRequiredTemplateFields = 76

Missing required template fields

MissingTemplateFieldValue = 77

Missing a required field value

MissingText = 78

The ad must have text.

MissingVisibleUrl = 79

A visible URL must be entered.

MissingWidth = 80

A width must be entered.

MultipleDistinctFeedsUnsupported = 81

Only 1 feed can be used as the source of ad customizer substitutions in a single ad.

MustUseTempAdUnionIdOnAdd = 82

TempAdUnionId must be use when adding template ads.

TooLong = 83

The string has too many characters.

TooShort = 84

The string has too few characters.

UnionDimensionsCannotChange = 85

Ad union dimensions cannot change for saved ads.

UnknownAddressComponent = 86

Address component is not {country, lat, lng}.

UnknownFieldName = 87

Unknown unique field name

UnknownUniqueName = 88

Unknown unique name (template element type specifier)

UnsupportedDimensions = 89

Unsupported ad dimension

UrlInvalidScheme = 90

URL starts with an invalid scheme.

UrlInvalidTopLevelDomain = 91

URL ends with an invalid top-level domain name.

UrlMalformed = 92

URL contains illegal characters.

UrlNoHost = 93

URL must contain a host name.

UrlNotEquivalent = 94

URL not equivalent during upgrade.

UrlHostNameTooLong = 95

URL host name too long to be stored as visible URL (applies to Ad Exchange ads)

UrlNoScheme = 96

URL must start with a scheme.

UrlNoTopLevelDomain = 97

URL should end in a valid domain extension, such as .com or .net.

UrlPathNotAllowed = 98

URL must not end with a path.

UrlPortNotAllowed = 99

URL must not specify a port.

UrlQueryNotAllowed = 100

URL must not contain a query.

UrlSchemeBeforeExpandedDynamicSearchAdTag = 102

A url scheme is not allowed in front of tag in tracking url template (for example, http://{lpurl})

UserDoesNotHaveAccessToTemplate = 103

The user does not have permissions to create a template ad for the given template.

InconsistentExpandableSettings = 104

Expandable setting is inconsistent/wrong. For example, an AdX ad is invalid if it has a expandable vendor format but no expanding directions specified, or expanding directions is specified, but the vendor format is not expandable.

InvalidFormat = 105

Format is invalid

InvalidFieldText = 106

The text of this field did not match a pattern of allowed values.

ElementNotPresent = 107

Template element is mising

ImageError = 108

Error occurred during image processing

ValueNotInRange = 109

The value is not within the valid range

FieldNotPresent = 110

Template element field is not present

AddressNotComplete = 111

Address is incomplete

AddressInvalid = 112

Invalid address

VideoRetrievalError = 113

Error retrieving specified video

AudioError = 114

Error processing audio

InvalidYoutubeDisplayUrl = 115

Display URL is incorrect for YouTube PYV ads

TooManyProductImages = 116

Too many product Images in GmailAd

TooManyProductVideos = 117

Too many product Videos in GmailAd

IncompatibleAdTypeAndDevicePreference = 118

The device preference is not compatible with the ad type

CalltrackingNotSupportedForCountry = 119

Call tracking is not supported for specified country.

CarrierSpecificShortNumberNotAllowed = 120

Carrier specific short number is not allowed.

DisallowedNumberType = 121

Specified phone number type is disallowed.

PhoneNumberNotSupportedForCountry = 122

Phone number not supported for country.

PhoneNumberNotSupportedWithCalltrackingForCountry = 123

Phone number not supported with call tracking enabled for country.

PremiumRateNumberNotAllowed = 124

Premium rate phone number is not allowed.

VanityPhoneNumberNotAllowed = 125

Vanity phone number is not allowed.

InvalidCallConversionTypeId = 126

Invalid call conversion type id.

CannotDisableCallConversionAndSetConversionTypeId = 127

Cannot disable call conversion and set conversion type id.

CannotSetPath2WithoutPath1 = 128

Cannot set path2 without path1.

MissingDynamicSearchAdsSettingDomainName = 129

Missing domain name in campaign setting when adding expanded dynamic search ad.

IncompatibleWithRestrictionType = 130

The associated ad is not compatible with restriction type.

CustomerConsentForCallRecordingRequired = 131

Consent for call recording is required for creating/updating call only ads. See https://support.google.com/google-ads/answer/7412639.

MissingImageOrMediaBundle = 132

Either an image or a media bundle is required in a display upload ad.

ProductTypeNotSupportedInThisCampaign = 133

The display upload product type is not supported in this campaign.

PlaceholderCannotHaveEmptyDefaultValue = 134

The default value of an ad placeholder can not be the empty string.

PlaceholderCountdownFunctionCannotHaveDefaultValue = 135

Ad placeholders with countdown functions must not have a default value.

PlaceholderDefaultValueMissing = 136

A previous ad placeholder that had a default value was found which means that all (non-countdown) placeholders must have a default value. This ad placeholder does not have a default value.

UnexpectedPlaceholderDefaultValue = 137

A previous ad placeholder that did not have a default value was found which means that no placeholders may have a default value. This ad placeholder does have a default value.

AdCustomizersMayNotBeAdjacent = 138

Two ad customizers may not be directly adjacent in an ad text. They must be separated by at least one character.

UpdatingAdWithNoEnabledAssociation = 139

The ad is not associated with any enabled AdGroupAd, and cannot be updated.

CallAdVerificationUrlFinalUrlDoesNotHaveSameDomain = 140

Call Ad verification url and final url don’t have same domain.

CallAdFinalUrlAndVerificationUrlCannotBothBeEmpty = 154

Final url and verification url cannot both be empty for call ads.

TooManyAdCustomizers = 141

Too many ad customizers in one asset.

InvalidAdCustomizerFormat = 142

The ad customizer tag is recognized, but the format is invalid.

NestedAdCustomizerSyntax = 143

Customizer tags cannot be nested.

UnsupportedAdCustomizerSyntax = 144

The ad customizer syntax used in the ad is not supported.

UnpairedBraceInAdCustomizerTag = 145

There exists unpaired brace in the ad customizer tag.

MoreThanOneCountdownTagTypeExists = 146

More than one type of countdown tag exists among all text lines.

DateTimeInCountdownTagIsInvalid = 147

Date time in the countdown tag is invalid.

DateTimeInCountdownTagIsPast = 148

Date time in the countdown tag is in the past.

UnrecognizedAdCustomizerTagFound = 149

Cannot recognize the ad customizer tag.

CustomizerTypeForbiddenForField = 150

Customizer type forbidden for this field.

InvalidCustomizerAttributeName = 151

Customizer attribute name is invalid.

StoreMismatch = 152

App store value does not match the value of the app store in the app specified in the campaign.

MissingRequiredImageAspectRatio = 153

Missing required image aspect ratio.

MismatchedAspectRatios = 155

Aspect ratios mismatch between different assets.

DuplicateImageAcrossCarouselCards = 156

Images must be unique between different carousel card assets.

Implementations

impl AdError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of AdError.

pub fn from_i32(value: i32) -> Option<AdError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a AdError, or None if value is not a valid variant.

impl AdError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.