Enum google_api_proto::google::ads::googleads::v15::errors::asset_group_error_enum::AssetGroupError

#[repr(i32)]
pub enum AssetGroupError {

    Unspecified = 0,
    Unknown = 1,
    DuplicateName = 2,
    CannotAddAssetGroupForCampaignType = 3,
    NotEnoughHeadlineAsset = 4,
    NotEnoughLongHeadlineAsset = 5,
    NotEnoughDescriptionAsset = 6,
    NotEnoughBusinessNameAsset = 7,
    NotEnoughMarketingImageAsset = 8,
    NotEnoughSquareMarketingImageAsset = 9,
    NotEnoughLogoAsset = 10,
    FinalUrlShoppingMerchantHomePageUrlDomainsDiffer = 11,
    Path1RequiredWhenPath2IsSet = 12,
    ShortDescriptionRequired = 13,
    FinalUrlRequired = 14,
    FinalUrlContainsInvalidDomainName = 15,
    AdCustomizerNotSupported = 16,
    CannotMutateAssetGroupForRemovedCampaign = 17,

}

Enum describing possible asset group errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

DuplicateName = 2

Each asset group in a single campaign must have a unique name.

CannotAddAssetGroupForCampaignType = 3

Cannot add asset group for the campaign type.

NotEnoughHeadlineAsset = 4

Not enough headline asset for a valid asset group.

NotEnoughLongHeadlineAsset = 5

Not enough long headline asset for a valid asset group.

NotEnoughDescriptionAsset = 6

Not enough description headline asset for a valid asset group.

NotEnoughBusinessNameAsset = 7

Not enough business name asset for a valid asset group.

NotEnoughMarketingImageAsset = 8

Not enough marketing image asset for a valid asset group.

NotEnoughSquareMarketingImageAsset = 9

Not enough square marketing image asset for a valid asset group.

NotEnoughLogoAsset = 10

Not enough logo asset for a valid asset group.

FinalUrlShoppingMerchantHomePageUrlDomainsDiffer = 11

Final url and shopping merchant url does not have the same domain.

Path1RequiredWhenPath2IsSet = 12

Path1 required when path2 is set.

ShortDescriptionRequired = 13

At least one short description asset is required for a valid asset group.

FinalUrlRequired = 14

Final url field is required for asset group.

FinalUrlContainsInvalidDomainName = 15

Final url contains invalid domain name.

AdCustomizerNotSupported = 16

Ad customizers are not supported in asset group’s text field.

CannotMutateAssetGroupForRemovedCampaign = 17

Cannot mutate asset group for campaign with removed status.

Implementations

impl AssetGroupError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of AssetGroupError.

pub fn from_i32(value: i32) -> Option<AssetGroupError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a AssetGroupError, or None if value is not a valid variant.

impl AssetGroupError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.