Enum google_api_proto::google::ads::googleads::v15::errors::conversion_upload_error_enum::ConversionUploadError

#[repr(i32)]
pub enum ConversionUploadError {

    Unspecified = 0,
    Unknown = 1,
    TooManyConversionsInRequest = 2,
    UnparseableGclid = 3,
    ConversionPrecedesEvent = 42,
    ExpiredEvent = 43,
    TooRecentEvent = 44,
    EventNotFound = 45,
    UnauthorizedCustomer = 8,
    TooRecentConversionAction = 10,
    ConversionTrackingNotEnabledAtImpressionTime = 11,
    ExternalAttributionDataSetForNonExternallyAttributedConversionAction = 12,
    ExternalAttributionDataNotSetForExternallyAttributedConversionAction = 13,
    OrderIdNotPermittedForExternallyAttributedConversionAction = 14,
    OrderIdAlreadyInUse = 15,
    DuplicateOrderId = 16,
    TooRecentCall = 17,
    ExpiredCall = 18,
    CallNotFound = 19,
    ConversionPrecedesCall = 20,
    ConversionTrackingNotEnabledAtCallTime = 21,
    UnparseableCallersPhoneNumber = 22,
    ClickConversionAlreadyExists = 23,
    CallConversionAlreadyExists = 24,
    DuplicateClickConversionInRequest = 25,
    DuplicateCallConversionInRequest = 26,
    CustomVariableNotEnabled = 28,
    CustomVariableValueContainsPii = 29,
    InvalidCustomerForClick = 30,
    InvalidCustomerForCall = 31,
    ConversionNotCompliantWithAttPolicy = 32,
    ClickNotFound = 33,
    InvalidUserIdentifier = 34,
    ExternallyAttributedConversionActionNotPermittedWithUserIdentifier = 35,
    UnsupportedUserIdentifier = 36,
    GbraidWbraidBothSet = 38,
    UnparseableWbraid = 39,
    UnparseableGbraid = 40,
    OnePerClickConversionActionNotPermittedWithBraid = 46,
    CustomerDataPolicyProhibitsEnhancedConversions = 47,
    CustomerNotAcceptedCustomerDataTerms = 48,
    OrderIdContainsPii = 49,
    CustomerNotEnabledEnhancedConversionsForLeads = 50,
    InvalidJobId = 52,
    NoConversionActionFound = 53,
    InvalidConversionActionType = 54,

}

Enum describing possible conversion upload errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

TooManyConversionsInRequest = 2

Upload fewer than 2001 events in a single request.

UnparseableGclid = 3

The imported gclid could not be decoded. Make sure you have not modified the click IDs.

ConversionPrecedesEvent = 42

The imported event has a conversion_date_time that precedes the click. Make sure your conversion_date_time is correct and try again.

ExpiredEvent = 43

The imported event can’t be recorded because its click occurred before this conversion’s click-through window. Make sure you import the most recent data.

TooRecentEvent = 44

The click associated with the given identifier or iOS URL parameter occurred less than 6 hours ago. Retry after 6 hours have passed.

EventNotFound = 45

The imported event could not be attributed to a click. This may be because the event did not come from a Google Ads campaign.

UnauthorizedCustomer = 8

The click ID or call is associated with an Ads account you don’t have access to. Make sure you import conversions for accounts managed by your manager account.

TooRecentConversionAction = 10

Can’t import events to a conversion action that was just created. Try importing again in 6 hours.

ConversionTrackingNotEnabledAtImpressionTime = 11

At the time of the click, conversion tracking was not enabled in the effective conversion account of the click’s Google Ads account.

ExternalAttributionDataSetForNonExternallyAttributedConversionAction = 12

The imported event includes external attribution data, but the conversion action isn’t set up to use an external attribution model. Make sure the conversion action is correctly configured and try again.

ExternalAttributionDataNotSetForExternallyAttributedConversionAction = 13

The conversion action is set up to use an external attribution model, but the imported event is missing data. Make sure imported events include the external attribution credit and all necessary fields.

OrderIdNotPermittedForExternallyAttributedConversionAction = 14

Order IDs can’t be used for a conversion measured with an external attribution model. Make sure the conversion is correctly configured and imported events include only necessary data and try again.

OrderIdAlreadyInUse = 15

The imported event includes an order ID that was previously recorded, so the event was not processed.

DuplicateOrderId = 16

Imported events include multiple conversions with the same order ID and were not processed. Make sure order IDs are unique and try again.

TooRecentCall = 17

Can’t import calls that occurred less than 6 hours ago. Try uploading again in 6 hours.

ExpiredCall = 18

The call can’t be recorded because it occurred before this conversion action’s lookback window. Make sure your import is configured to get the most recent data.

CallNotFound = 19

The call or click leading to the imported event can’t be found. Make sure your data source is set up to include correct identifiers.

ConversionPrecedesCall = 20

The call has a conversion_date_time that precedes the associated click. Make sure your conversion_date_time is correct.

ConversionTrackingNotEnabledAtCallTime = 21

At the time of the imported call, conversion tracking was not enabled in the effective conversion account of the click’s Google Ads account.

UnparseableCallersPhoneNumber = 22

Make sure phone numbers are formatted as E.164 (+16502531234), International (+64 3-331 6005), or US national number (6502531234).

ClickConversionAlreadyExists = 23

The imported event has the same click and conversion_date_time as an existing conversion. Use a unique conversion_date_time or order ID for each unique event and try again.

CallConversionAlreadyExists = 24

The imported call has the same conversion_date_time as an existing conversion. Make sure your conversion_date_time correctly configured and try again.

DuplicateClickConversionInRequest = 25

Multiple events have the same click and conversion_date_time. Make sure your conversion_date_time is correctly configured and try again.

DuplicateCallConversionInRequest = 26

Multiple events have the same call and conversion_date_time. Make sure your conversion_date_time is correctly configured and try again.

CustomVariableNotEnabled = 28

Enable the custom variable in your conversion settings and try again.

CustomVariableValueContainsPii = 29

Can’t import events with custom variables containing personally-identifiable information (PII). Remove these variables and try again.

InvalidCustomerForClick = 30

The click from the imported event is associated with a different Google Ads account. Make sure you’re importing to the correct account.

InvalidCustomerForCall = 31

The click from the call is associated with a different Google Ads account. Make sure you’re importing to the correct account. Query conversion_tracking_setting.google_ads_conversion_customer on Customer to identify the correct account.

ConversionNotCompliantWithAttPolicy = 32

The connversion can’t be imported because the conversion source didn’t comply with Apple App Transparency Tracking (ATT) policies or because the customer didn’t consent to tracking.

ClickNotFound = 33

The email address or phone number for this event can’t be matched to a click. This may be because it didn’t come from a Google Ads campaign, and you can safely ignore this warning. If this includes more imported events than is expected, you may need to check your setup.

InvalidUserIdentifier = 34

Make sure you hash user provided data using SHA-256 and ensure you are normalizing according to the guidelines.

ExternallyAttributedConversionActionNotPermittedWithUserIdentifier = 35

User provided data can’t be used with external attribution models. Use a different attribution model or omit user identifiers and try again.

UnsupportedUserIdentifier = 36

The provided user identifiers are not supported. Use only hashed email or phone number and try again.

GbraidWbraidBothSet = 38

Can’t use both gbraid and wbraid parameters. Use only 1 and try again.

UnparseableWbraid = 39

Can’t parse event import data. Check if your wbraid parameter was not modified and try again.

UnparseableGbraid = 40

Can’t parse event import data. Check if your gbraid parameter was not modified and try again.

OnePerClickConversionActionNotPermittedWithBraid = 46

Conversion actions that use one-per-click counting can’t be used with gbraid or wbraid parameters.

CustomerDataPolicyProhibitsEnhancedConversions = 47

Enhanced conversions can’t be used for this account because of Google customer data policies. Contact your Google representative.

CustomerNotAcceptedCustomerDataTerms = 48

Make sure you agree to the customer data processing terms in conversion settings and try again. You can check your setting by querying conversion_tracking_setting.accepted_customer_data_terms on Customer.

OrderIdContainsPii = 49

Can’t import events with order IDs containing personally-identifiable information (PII).

CustomerNotEnabledEnhancedConversionsForLeads = 50

Make sure you’ve turned on enhanced conversions for leads in conversion settings and try again. You can check your setting by querying conversion_tracking_setting.enhanced_conversions_for_leads_enabled on Customer.

InvalidJobId = 52

The provided job id in the request is not within the allowed range. A job ID must be a positive integer in the range [1, 2^31).

NoConversionActionFound = 53

The conversion action specified in the upload request cannot be found. Make sure it’s available in this account.

InvalidConversionActionType = 54

The conversion action specified in the upload request isn’t set up for uploading conversions.

Implementations

impl ConversionUploadError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionUploadError.

pub fn from_i32(value: i32) -> Option<ConversionUploadError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ConversionUploadError, or None if value is not a valid variant.

impl ConversionUploadError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.