Enum google_api_proto::google::ads::googleads::v15::errors::feed_item_validation_error_enum::FeedItemValidationError

#[repr(i32)]
pub enum FeedItemValidationError {

    Unspecified = 0,
    Unknown = 1,
    StringTooShort = 2,
    StringTooLong = 3,
    ValueNotSpecified = 4,
    InvalidDomesticPhoneNumberFormat = 5,
    InvalidPhoneNumber = 6,
    PhoneNumberNotSupportedForCountry = 7,
    PremiumRateNumberNotAllowed = 8,
    DisallowedNumberType = 9,
    ValueOutOfRange = 10,
    CalltrackingNotSupportedForCountry = 11,
    CustomerNotInAllowlistForCalltracking = 99,
    InvalidCountryCode = 13,
    InvalidAppId = 14,
    MissingAttributesForFields = 15,
    InvalidTypeId = 16,
    InvalidEmailAddress = 17,
    InvalidHttpsUrl = 18,
    MissingDeliveryAddress = 19,
    StartDateAfterEndDate = 20,
    MissingFeedItemStartTime = 21,
    MissingFeedItemEndTime = 22,
    MissingFeedItemId = 23,
    VanityPhoneNumberNotAllowed = 24,
    InvalidReviewExtensionSnippet = 25,
    InvalidNumberFormat = 26,
    InvalidDateFormat = 27,
    InvalidPriceFormat = 28,
    UnknownPlaceholderField = 29,
    MissingEnhancedSitelinkDescriptionLine = 30,
    ReviewExtensionSourceIneligible = 31,
    HyphensInReviewExtensionSnippet = 32,
    DoubleQuotesInReviewExtensionSnippet = 33,
    QuotesInReviewExtensionSnippet = 34,
    InvalidFormEncodedParams = 35,
    InvalidUrlParameterName = 36,
    NoGeocodingResult = 37,
    SourceNameInReviewExtensionText = 38,
    CarrierSpecificShortNumberNotAllowed = 39,
    InvalidPlaceholderFieldId = 40,
    InvalidUrlTag = 41,
    ListTooLong = 42,
    InvalidAttributesCombination = 43,
    DuplicateValues = 44,
    InvalidCallConversionActionId = 45,
    CannotSetWithoutFinalUrls = 46,
    AppIdDoesntExistInAppStore = 47,
    InvalidFinalUrl = 48,
    InvalidTrackingUrl = 49,
    InvalidFinalUrlForAppDownloadUrl = 50,
    ListTooShort = 51,
    InvalidUserAction = 52,
    InvalidTypeName = 53,
    InvalidEventChangeStatus = 54,
    InvalidSnippetsHeader = 55,
    InvalidAndroidAppLink = 56,
    NumberTypeWithCalltrackingNotSupportedForCountry = 57,
    ReservedKeywordOther = 58,
    DuplicateOptionLabels = 59,
    DuplicateOptionPrefills = 60,
    UnequalListLengths = 61,
    InconsistentCurrencyCodes = 62,
    PriceExtensionHasDuplicatedHeaders = 63,
    ItemHasDuplicatedHeaderAndDescription = 64,
    PriceExtensionHasTooFewItems = 65,
    UnsupportedValue = 66,
    InvalidFinalMobileUrl = 67,
    InvalidKeywordlessAdRuleLabel = 68,
    ValueTrackParameterNotSupported = 69,
    UnsupportedValueInSelectedLanguage = 70,
    InvalidIosAppLink = 71,
    MissingIosAppLinkOrIosAppStoreId = 72,
    PromotionInvalidTime = 73,
    PromotionCannotSetPercentOffAndMoneyAmountOff = 74,
    PromotionCannotSetPromotionCodeAndOrdersOverAmount = 75,
    TooManyDecimalPlacesSpecified = 76,
    AdCustomizersNotAllowed = 77,
    InvalidLanguageCode = 78,
    UnsupportedLanguage = 79,
    IfFunctionNotAllowed = 80,
    InvalidFinalUrlSuffix = 81,
    InvalidTagInFinalUrlSuffix = 82,
    InvalidFinalUrlSuffixFormat = 83,
    CustomerConsentForCallRecordingRequired = 84,
    OnlyOneDeliveryOptionIsAllowed = 85,
    NoDeliveryOptionIsSet = 86,
    InvalidConversionReportingState = 87,
    ImageSizeWrong = 88,
    EmailDeliveryNotAvailableInCountry = 89,
    AutoReplyNotAvailableInCountry = 90,
    InvalidLatitudeValue = 91,
    InvalidLongitudeValue = 92,
    TooManyLabels = 93,
    InvalidImageUrl = 94,
    MissingLatitudeValue = 95,
    MissingLongitudeValue = 96,
    AddressNotFound = 97,
    AddressNotTargetable = 98,
    InvalidAssetId = 100,
    IncompatibleAssetType = 101,
    ImageErrorUnexpectedSize = 102,
    ImageErrorAspectRatioNotAllowed = 103,
    ImageErrorFileTooLarge = 104,
    ImageErrorFormatNotAllowed = 105,
    ImageErrorConstraintsViolated = 106,
    ImageErrorServerError = 107,

}

The possible validation errors of a feed item.

Variants

Unspecified = 0

No value has been specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

StringTooShort = 2

String is too short.

StringTooLong = 3

String is too long.

ValueNotSpecified = 4

Value is not provided.

InvalidDomesticPhoneNumberFormat = 5

Phone number format is invalid for region.

InvalidPhoneNumber = 6

String does not represent a phone number.

PhoneNumberNotSupportedForCountry = 7

Phone number format is not compatible with country code.

PremiumRateNumberNotAllowed = 8

Premium rate number is not allowed.

DisallowedNumberType = 9

Phone number type is not allowed.

ValueOutOfRange = 10

Specified value is outside of the valid range.

CalltrackingNotSupportedForCountry = 11

Call tracking is not supported in the selected country.

CustomerNotInAllowlistForCalltracking = 99

Customer is not on the allow-list for call tracking.

InvalidCountryCode = 13

Country code is invalid.

InvalidAppId = 14

The specified mobile app id is invalid.

MissingAttributesForFields = 15

Some required field attributes are missing.

InvalidTypeId = 16

Invalid email button type for email extension.

InvalidEmailAddress = 17

Email address is invalid.

InvalidHttpsUrl = 18

The HTTPS URL in email extension is invalid.

MissingDeliveryAddress = 19

Delivery address is missing from email extension.

StartDateAfterEndDate = 20

FeedItem scheduling start date comes after end date.

MissingFeedItemStartTime = 21

FeedItem scheduling start time is missing.

MissingFeedItemEndTime = 22

FeedItem scheduling end time is missing.

MissingFeedItemId = 23

Cannot compute system attributes on a FeedItem that has no FeedItemId.

VanityPhoneNumberNotAllowed = 24

Call extension vanity phone numbers are not supported.

InvalidReviewExtensionSnippet = 25

Invalid review text.

InvalidNumberFormat = 26

Invalid format for numeric value in ad parameter.

InvalidDateFormat = 27

Invalid format for date value in ad parameter.

InvalidPriceFormat = 28

Invalid format for price value in ad parameter.

UnknownPlaceholderField = 29

Unrecognized type given for value in ad parameter.

MissingEnhancedSitelinkDescriptionLine = 30

Enhanced sitelinks must have both description lines specified.

ReviewExtensionSourceIneligible = 31

Review source is ineligible.

HyphensInReviewExtensionSnippet = 32

Review text cannot contain hyphens or dashes.

DoubleQuotesInReviewExtensionSnippet = 33

Review text cannot contain double quote characters.

QuotesInReviewExtensionSnippet = 34

Review text cannot contain quote characters.

InvalidFormEncodedParams = 35

Parameters are encoded in the wrong format.

InvalidUrlParameterName = 36

URL parameter name must contain only letters, numbers, underscores, and dashes.

NoGeocodingResult = 37

Cannot find address location.

SourceNameInReviewExtensionText = 38

Review extension text has source name.

CarrierSpecificShortNumberNotAllowed = 39

Some phone numbers can be shorter than usual. Some of these short numbers are carrier-specific, and we disallow those in ad extensions because they will not be available to all users.

InvalidPlaceholderFieldId = 40

Triggered when a request references a placeholder field id that does not exist.

InvalidUrlTag = 41

URL contains invalid ValueTrack tags or format.

ListTooLong = 42

Provided list exceeds acceptable size.

InvalidAttributesCombination = 43

Certain combinations of attributes aren’t allowed to be specified in the same feed item.

DuplicateValues = 44

An attribute has the same value repeatedly.

InvalidCallConversionActionId = 45

Advertisers can link a conversion action with a phone number to indicate that sufficiently long calls forwarded to that phone number should be counted as conversions of the specified type. This is an error message indicating that the conversion action specified is invalid (for example, the conversion action does not exist within the appropriate Google Ads account, or it is a type of conversion not appropriate to phone call conversions).

CannotSetWithoutFinalUrls = 46

Tracking template requires final url to be set.

AppIdDoesntExistInAppStore = 47

An app id was provided that doesn’t exist in the given app store.

InvalidFinalUrl = 48

Invalid U2 final url.

InvalidTrackingUrl = 49

Invalid U2 tracking url.

InvalidFinalUrlForAppDownloadUrl = 50

Final URL should start from App download URL.

ListTooShort = 51

List provided is too short.

InvalidUserAction = 52

User Action field has invalid value.

InvalidTypeName = 53

Type field has invalid value.

InvalidEventChangeStatus = 54

Change status for event is invalid.

InvalidSnippetsHeader = 55

The header of a structured snippets extension is not one of the valid headers.

InvalidAndroidAppLink = 56

Android app link is not formatted correctly

NumberTypeWithCalltrackingNotSupportedForCountry = 57

Phone number incompatible with call tracking for country.

ReservedKeywordOther = 58

The input is identical to a reserved keyword

DuplicateOptionLabels = 59

Each option label in the message extension must be unique.

DuplicateOptionPrefills = 60

Each option prefill in the message extension must be unique.

UnequalListLengths = 61

In message extensions, the number of optional labels and optional prefills must be the same.

InconsistentCurrencyCodes = 62

All currency codes in an ad extension must be the same.

PriceExtensionHasDuplicatedHeaders = 63

Headers in price extension are not unique.

ItemHasDuplicatedHeaderAndDescription = 64

Header and description in an item are the same.

PriceExtensionHasTooFewItems = 65

Price extension has too few items.

UnsupportedValue = 66

The given value is not supported.

InvalidFinalMobileUrl = 67

Invalid final mobile url.

InvalidKeywordlessAdRuleLabel = 68

The given string value of Label contains invalid characters

ValueTrackParameterNotSupported = 69

The given URL contains value track parameters.

UnsupportedValueInSelectedLanguage = 70

The given value is not supported in the selected language of an extension.

InvalidIosAppLink = 71

The iOS app link is not formatted correctly.

MissingIosAppLinkOrIosAppStoreId = 72

iOS app link or iOS app store id is missing.

PromotionInvalidTime = 73

Promotion time is invalid.

PromotionCannotSetPercentOffAndMoneyAmountOff = 74

Both the percent off and money amount off fields are set.

PromotionCannotSetPromotionCodeAndOrdersOverAmount = 75

Both the promotion code and orders over amount fields are set.

TooManyDecimalPlacesSpecified = 76

Too many decimal places are specified.

AdCustomizersNotAllowed = 77

Ad Customizers are present and not allowed.

InvalidLanguageCode = 78

Language code is not valid.

UnsupportedLanguage = 79

Language is not supported.

IfFunctionNotAllowed = 80

IF Function is present and not allowed.

InvalidFinalUrlSuffix = 81

Final url suffix is not valid.

InvalidTagInFinalUrlSuffix = 82

Final url suffix contains an invalid tag.

InvalidFinalUrlSuffixFormat = 83

Final url suffix is formatted incorrectly.

CustomerConsentForCallRecordingRequired = 84

Consent for call recording, which is required for the use of call extensions, was not provided by the advertiser. See https://support.google.com/google-ads/answer/7412639.

OnlyOneDeliveryOptionIsAllowed = 85

Multiple message delivery options are set.

NoDeliveryOptionIsSet = 86

No message delivery option is set.

InvalidConversionReportingState = 87

String value of conversion reporting state field is not valid.

ImageSizeWrong = 88

Image size is not right.

EmailDeliveryNotAvailableInCountry = 89

Email delivery is not supported in the country specified in the country code field.

AutoReplyNotAvailableInCountry = 90

Auto reply is not supported in the country specified in the country code field.

InvalidLatitudeValue = 91

Invalid value specified for latitude.

InvalidLongitudeValue = 92

Invalid value specified for longitude.

TooManyLabels = 93

Too many label fields provided.

InvalidImageUrl = 94

Invalid image url.

MissingLatitudeValue = 95

Latitude value is missing.

MissingLongitudeValue = 96

Longitude value is missing.

AddressNotFound = 97

Unable to find address.

AddressNotTargetable = 98

Cannot target provided address.

InvalidAssetId = 100

The specified asset ID does not exist.

IncompatibleAssetType = 101

The asset type cannot be set for the field.

ImageErrorUnexpectedSize = 102

The image has unexpected size.

ImageErrorAspectRatioNotAllowed = 103

The image aspect ratio is not allowed.

ImageErrorFileTooLarge = 104

The image file is too large.

ImageErrorFormatNotAllowed = 105

The image format is unsupported.

ImageErrorConstraintsViolated = 106

Image violates constraints without more details.

ImageErrorServerError = 107

An error occurred when validating image.

Implementations

impl FeedItemValidationError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of FeedItemValidationError.

pub fn from_i32(value: i32) -> Option<FeedItemValidationError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a FeedItemValidationError, or None if value is not a valid variant.

impl FeedItemValidationError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.