Enum google_api_proto::google::ads::googleads::v15::errors::image_error_enum::ImageError

#[repr(i32)]
pub enum ImageError {

    Unspecified = 0,
    Unknown = 1,
    InvalidImage = 2,
    StorageError = 3,
    BadRequest = 4,
    UnexpectedSize = 5,
    AnimatedNotAllowed = 6,
    AnimationTooLong = 7,
    ServerError = 8,
    CmykJpegNotAllowed = 9,
    FlashNotAllowed = 10,
    FlashWithoutClicktag = 11,
    FlashErrorAfterFixingClickTag = 12,
    AnimatedVisualEffect = 13,
    FlashError = 14,
    LayoutProblem = 15,
    ProblemReadingImageFile = 16,
    ErrorStoringImage = 17,
    AspectRatioNotAllowed = 18,
    FlashHasNetworkObjects = 19,
    FlashHasNetworkMethods = 20,
    FlashHasUrl = 21,
    FlashHasMouseTracking = 22,
    FlashHasRandomNum = 23,
    FlashSelfTargets = 24,
    FlashBadGeturlTarget = 25,
    FlashVersionNotSupported = 26,
    FlashWithoutHardCodedClickUrl = 27,
    InvalidFlashFile = 28,
    FailedToFixClickTagInFlash = 29,
    FlashAccessesNetworkResources = 30,
    FlashExternalJsCall = 31,
    FlashExternalFsCall = 32,
    FileTooLarge = 33,
    ImageDataTooLarge = 34,
    ImageProcessingError = 35,
    ImageTooSmall = 36,
    InvalidInput = 37,
    ProblemReadingFile = 38,
    ImageConstraintsViolated = 39,
    FormatNotAllowed = 40,

}

Enum describing possible image errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

InvalidImage = 2

The image is not valid.

StorageError = 3

The image could not be stored.

BadRequest = 4

There was a problem with the request.

UnexpectedSize = 5

The image is not of legal dimensions.

AnimatedNotAllowed = 6

Animated image are not permitted.

AnimationTooLong = 7

Animation is too long.

ServerError = 8

There was an error on the server.

CmykJpegNotAllowed = 9

Image cannot be in CMYK color format.

FlashNotAllowed = 10

Flash images are not permitted.

FlashWithoutClicktag = 11

Flash images must support clickTag.

FlashErrorAfterFixingClickTag = 12

A flash error has occurred after fixing the click tag.

AnimatedVisualEffect = 13

Unacceptable visual effects.

FlashError = 14

There was a problem with the flash image.

LayoutProblem = 15

Incorrect image layout.

ProblemReadingImageFile = 16

There was a problem reading the image file.

ErrorStoringImage = 17

There was an error storing the image.

AspectRatioNotAllowed = 18

The aspect ratio of the image is not allowed.

FlashHasNetworkObjects = 19

Flash cannot have network objects.

FlashHasNetworkMethods = 20

Flash cannot have network methods.

FlashHasUrl = 21

Flash cannot have a Url.

FlashHasMouseTracking = 22

Flash cannot use mouse tracking.

FlashHasRandomNum = 23

Flash cannot have a random number.

FlashSelfTargets = 24

Ad click target cannot be ‘_self’.

FlashBadGeturlTarget = 25

GetUrl method should only use ‘_blank’.

FlashVersionNotSupported = 26

Flash version is not supported.

FlashWithoutHardCodedClickUrl = 27

Flash movies need to have hard coded click URL or clickTAG

InvalidFlashFile = 28

Uploaded flash file is corrupted.

FailedToFixClickTagInFlash = 29

Uploaded flash file can be parsed, but the click tag can not be fixed properly.

FlashAccessesNetworkResources = 30

Flash movie accesses network resources

FlashExternalJsCall = 31

Flash movie attempts to call external javascript code

FlashExternalFsCall = 32

Flash movie attempts to call flash system commands

FileTooLarge = 33

Image file is too large.

ImageDataTooLarge = 34

Image data is too large.

ImageProcessingError = 35

Error while processing the image.

ImageTooSmall = 36

Image is too small.

InvalidInput = 37

Input was invalid.

ProblemReadingFile = 38

There was a problem reading the image file.

ImageConstraintsViolated = 39

Image constraints are violated, but details like ASPECT_RATIO_NOT_ALLOWED can’t be provided. This happens when asset spec contains more than one constraint and different criteria of different constraints are violated.

FormatNotAllowed = 40

Image format is not allowed.

Implementations

impl ImageError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ImageError.

pub fn from_i32(value: i32) -> Option<ImageError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ImageError, or None if value is not a valid variant.

impl ImageError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.