Module google_api_proto::google::ads::googleads::v15::resources

Modules

accessible_bidding_strategy

Nested message and enum types in AccessibleBiddingStrategy.

account_budget

Nested message and enum types in AccountBudget.

account_budget_proposal

Nested message and enum types in AccountBudgetProposal.

account_link

Nested message and enum types in AccountLink.

ad

Nested message and enum types in Ad.

ad_group

Nested message and enum types in AdGroup.

ad_group_bid_modifier

Nested message and enum types in AdGroupBidModifier.

ad_group_criterion

Nested message and enum types in AdGroupCriterion.

ad_group_criterion_simulation

Nested message and enum types in AdGroupCriterionSimulation.

ad_group_simulation

Nested message and enum types in AdGroupSimulation.

asset

Nested message and enum types in Asset.

asset_group_signal

Nested message and enum types in AssetGroupSignal.

asset_set

Nested message and enum types in AssetSet.

attribute_field_mapping

Nested message and enum types in AttributeFieldMapping.

batch_job

Nested message and enum types in BatchJob.

bidding_strategy

Nested message and enum types in BiddingStrategy.

bidding_strategy_simulation

Nested message and enum types in BiddingStrategySimulation.

billing_setup

Nested message and enum types in BillingSetup.

campaign

Nested message and enum types in Campaign.

campaign_bid_modifier

Nested message and enum types in CampaignBidModifier.

campaign_criterion

Nested message and enum types in CampaignCriterion.

campaign_simulation

Nested message and enum types in CampaignSimulation.

change_event

Nested message and enum types in ChangeEvent.

conversion_action

Nested message and enum types in ConversionAction.

conversion_value_rule

Nested message and enum types in ConversionValueRule.

custom_audience_member

Nested message and enum types in CustomAudienceMember.

customer_lifecycle_goal

Nested message and enum types in CustomerLifecycleGoal.

customer_negative_criterion

Nested message and enum types in CustomerNegativeCriterion.

customer_sk_ad_network_conversion_value_schema

Nested message and enum types in CustomerSkAdNetworkConversionValueSchema.

extension_feed_item

Nested message and enum types in ExtensionFeedItem.

feed

Nested message and enum types in Feed.

feed_attribute_operation

Nested message and enum types in FeedAttributeOperation.

feed_item_set

Nested message and enum types in FeedItemSet.

feed_item_target

Nested message and enum types in FeedItemTarget.

feed_mapping

Nested message and enum types in FeedMapping.

invoice

Nested message and enum types in Invoice.

keyword_plan_forecast_period

Nested message and enum types in KeywordPlanForecastPeriod.

listing_group_filter_dimension

Nested message and enum types in ListingGroupFilterDimension.

local_services_verification_artifact

Nested message and enum types in LocalServicesVerificationArtifact.

media_file

Nested message and enum types in MediaFile.

offline_conversion_error

Nested message and enum types in OfflineConversionError.

offline_conversion_summary

Nested message and enum types in OfflineConversionSummary.

offline_user_data_job

Nested message and enum types in OfflineUserDataJob.

product_category_constant

Nested message and enum types in ProductCategoryConstant.

product_link

Nested message and enum types in ProductLink.

product_link_invitation

Nested message and enum types in ProductLinkInvitation.

recommendation

Nested message and enum types in Recommendation.

shared_criterion

Nested message and enum types in SharedCriterion.

smart_campaign_setting

Nested message and enum types in SmartCampaignSetting.

user_list

Nested message and enum types in UserList.

Structs

AccessibleBiddingStrategy

Represents a view of BiddingStrategies owned by and shared with the customer.

AccountBudget

An account-level budget. It contains information about the budget itself, as well as the most recently approved changes to the budget and proposed changes that are pending approval. The proposed changes that are pending approval, if any, are found in ‘pending_proposal’. Effective details about the budget are found in fields prefixed ‘approved_’, ‘adjusted_’ and those without a prefix. Since some effective details may differ from what the user had originally requested (for example, spending limit), these differences are juxtaposed through ‘proposed_’, ‘approved_’, and possibly ‘adjusted_’ fields.

AccountBudgetProposal

An account-level budget proposal.

AccountLink

Represents the data sharing connection between a Google Ads account and another account

Ad

An ad.

AdGroup

An ad group.

AdGroupAd

An ad group ad.

AdGroupAdAssetCombinationView

A view on the usage of ad group ad asset combination. Now we only support AdGroupAdAssetCombinationView for Responsive Search Ads, with more ad types planned for the future.

AdGroupAdAssetPolicySummary

Contains policy information for an ad group ad asset.

AdGroupAdAssetView

A link between an AdGroupAd and an Asset. Currently we only support AdGroupAdAssetView for AppAds and Responsive Search Ads.

AdGroupAdLabel

A relationship between an ad group ad and a label.

AdGroupAdPolicySummary

Contains policy information for an ad.

AdGroupAsset

A link between an ad group and an asset.

AdGroupAssetSet

AdGroupAssetSet is the linkage between an ad group and an asset set. Creating an AdGroupAssetSet links an asset set with an ad group.

AdGroupAudienceView

An ad group audience view. Includes performance data from interests and remarketing lists for Display Network and YouTube Network ads, and remarketing lists for search ads (RLSA), aggregated at the audience level.

AdGroupBidModifier

Represents an ad group bid modifier.

AdGroupCriterion

An ad group criterion.

AdGroupCriterionCustomizer

A customizer value for the associated CustomizerAttribute at the AdGroupCriterion level.

AdGroupCriterionLabel

A relationship between an ad group criterion and a label.

AdGroupCriterionSimulation

An ad group criterion simulation. Supported combinations of advertising channel type, criterion type, simulation type, and simulation modification method are detailed below respectively. Hotel AdGroupCriterion simulation operations starting in V5.

AdGroupCustomizer

A customizer value for the associated CustomizerAttribute at the AdGroup level.

AdGroupExtensionSetting

An ad group extension setting.

AdGroupFeed

An ad group feed.

AdGroupLabel

A relationship between an ad group and a label.

AdGroupSimulation

An ad group simulation. Supported combinations of advertising channel type, simulation type and simulation modification method is detailed below respectively.

AdParameter

An ad parameter that is used to update numeric values (such as prices or inventory levels) in any text line of an ad (including URLs). There can be a maximum of two AdParameters per ad group criterion. (One with parameter_index = 1 and one with parameter_index = 2.) In the ad the parameters are referenced by a placeholder of the form “{param#:value}”. For example, “{param1:$17}”

AdScheduleView

An ad schedule view summarizes the performance of campaigns by AdSchedule criteria.

AdvertisingPartnerLinkIdentifier

The identifier for the Advertising Partner Google Ads account.

AgeRangeView

An age range view.

AndroidPrivacySharedKeyGoogleAdGroup

An Android privacy shared key view for Google ad group key.

AndroidPrivacySharedKeyGoogleCampaign

An Android privacy shared key view for Google campaign key.

AndroidPrivacySharedKeyGoogleNetworkType

An Android privacy shared key view for Google network type key.

Asset

Asset is a part of an ad which can be shared across multiple ads. It can be an image (ImageAsset), a video (YoutubeVideoAsset), etc. Assets are immutable and cannot be removed. To stop an asset from serving, remove the asset from the entity that is using it.

AssetFieldTypePolicySummary

Contains policy information for an asset under AssetFieldType context.

AssetFieldTypeView

An asset field type view. This view reports non-overcounted metrics for each asset field type when the asset is used as extension.

AssetGroup

An asset group. AssetGroupAsset is used to link an asset to the asset group. AssetGroupSignal is used to associate a signal to an asset group.

AssetGroupAsset

AssetGroupAsset is the link between an asset and an asset group. Adding an AssetGroupAsset links an asset with an asset group.

AssetGroupAssetCombinationData

Asset group asset combination data

AssetGroupListingGroupFilter

AssetGroupListingGroupFilter represents a listing group filter tree node in an asset group.

AssetGroupProductGroupView

An asset group product group view.

AssetGroupSignal

AssetGroupSignal represents a signal in an asset group. The existence of a signal tells the performance max campaign who’s most likely to convert. Performance Max uses the signal to look for new people with similar or stronger intent to find conversions across Search, Display, Video, and more.

AssetGroupTopCombinationView

A view on the usage of ad group ad asset combination.

AssetPolicySummary

Contains policy information for an asset.

AssetSet

An asset set representing a collection of assets. Use AssetSetAsset to link an asset to the asset set.

AssetSetAsset

AssetSetAsset is the link between an asset and an asset set. Adding an AssetSetAsset links an asset with an asset set.

AssetSetTypeView

An asset set type view. This view reports non-overcounted metrics for each asset set type. Child asset set types are not included in this report. Their stats are aggregated under the parent asset set type.

AttributeFieldMapping

Maps from feed attribute id to a placeholder or criterion field id.

Audience

Audience is an effective targeting option that lets you intersect different segment attributes, such as detailed demographics and affinities, to create audiences that represent sections of your target segments.

BackgroundCheckVerificationArtifact

A proto holding information specific to local services background check.

BatchJob

A list of mutates being processed asynchronously. The mutates are uploaded by the user. The mutates themselves aren’t readable and the results of the job can only be read using BatchJobService.ListBatchJobResults.

BiddingDataExclusion

Represents a bidding data exclusion.

BiddingSeasonalityAdjustment

Represents a bidding seasonality adjustment.

BiddingStrategy

A bidding strategy.

BiddingStrategySimulation

A bidding strategy simulation. Supported combinations of simulation type and simulation modification method are detailed below respectively.

BillingSetup

A billing setup, which associates a payments account and an advertiser. A billing setup is specific to one advertiser.

CallReportingSetting

Call reporting setting for a customer. Only mutable in an update operation.

CallView

A call view that includes data for call tracking of call-only ads or call extensions.

Campaign

A campaign.

CampaignAsset

A link between a Campaign and an Asset.

CampaignAssetSet

CampaignAssetSet is the linkage between a campaign and an asset set. Adding a CampaignAssetSet links an asset set with a campaign.

CampaignAudienceView

A campaign audience view. Includes performance data from interests and remarketing lists for Display Network and YouTube Network ads, and remarketing lists for search ads (RLSA), aggregated by campaign and audience criterion. This view only includes audiences attached at the campaign level.

CampaignBidModifier

Represents a bid-modifiable only criterion at the campaign level.

CampaignBudget

A campaign budget.

CampaignConversionGoal

The biddability setting for the specified campaign only for all conversion actions with a matching category and origin.

CampaignCriterion

A campaign criterion.

CampaignCustomizer

A customizer value for the associated CustomizerAttribute at the Campaign level.

CampaignDraft

A campaign draft.

CampaignExtensionSetting

A campaign extension setting.

CampaignFeed

A campaign feed.

CampaignGroup

A campaign group.

CampaignLabel

Represents a relationship between a campaign and a label.

CampaignLifecycleGoal

Campaign level customer lifecycle goal settings.

CampaignSearchTermInsight

A Campaign search term view. Historical data is available starting March 2023.

CampaignSharedSet

CampaignSharedSets are used for managing the shared sets associated with a campaign.

CampaignSimulation

A campaign simulation. Supported combinations of advertising channel type, simulation type and simulation modification method is detailed below respectively.

CarrierConstant

A carrier criterion that can be used in campaign targeting.

ChangeEvent

Describes the granular change of returned resources of certain resource types. Changes made through the UI or API in the past 30 days are included. Previous and new values of the changed fields are shown. ChangeEvent could have up to 3 minutes delay to reflect a new change.

ChangeStatus

Describes the status of returned resource. ChangeStatus could have up to 3 minutes delay to reflect a new change.

ClickView

A click view with metrics aggregated at each click level, including both valid and invalid clicks. For non-Search campaigns, metrics.clicks represents the number of valid and invalid interactions. Queries including ClickView must have a filter limiting the results to one day and can be requested for dates back to 90 days before the time of the request.

CombinedAudience

Describe a resource for combined audiences which includes different audiences.

ContactDetails

Fields containing consumer contact details.

ConversionAction

A conversion action.

ConversionCustomVariable

A conversion custom variable See “About custom variables for conversions” at https://support.google.com/google-ads/answer/9964350

ConversionGoalCampaignConfig

Conversion goal settings for a Campaign.

ConversionTrackingSetting

A collection of customer-wide settings related to Google Ads Conversion Tracking.

ConversionValueRule

A conversion value rule

ConversionValueRuleSet

A conversion value rule set

CurrencyConstant

A currency constant.

CustomAudience

A custom audience. This is a list of users by interest.

CustomAudienceMember

A member of custom audience. A member can be a KEYWORD, URL, PLACE_CATEGORY or APP. It can only be created or removed but not changed.

CustomConversionGoal

Custom conversion goal that can make arbitrary conversion actions biddable.

CustomInterest

A custom interest. This is a list of users by interest.

CustomInterestMember

A member of custom interest audience. A member can be a keyword or url. It is immutable, that is, it can only be created or removed but not changed.

CustomLeadFormSubmissionField

Fields in the submitted custom question

Customer

A customer.

CustomerAcquisitionGoalSettings

The customer acquisition goal settings for the campaign.

CustomerAgreementSetting

Customer Agreement Setting for a customer.

CustomerAsset

A link between a customer and an asset.

CustomerAssetSet

CustomerAssetSet is the linkage between a customer and an asset set. Adding a CustomerAssetSet links an asset set with a customer.

CustomerClient

A link between the given customer and a client customer. CustomerClients only exist for manager customers. All direct and indirect client customers are included, as well as the manager itself.

CustomerClientLink

Represents customer client link relationship.

CustomerConversionGoal

Biddability control for conversion actions with a matching category and origin.

CustomerCustomizer

A customizer value for the associated CustomizerAttribute at the Customer level.

CustomerExtensionSetting

A customer extension setting.

CustomerFeed

A customer feed.

CustomerLabel

Represents a relationship between a customer and a label. This customer may not have access to all the labels attached to it. Additional CustomerLabels may be returned by increasing permissions with login-customer-id.

CustomerLifecycleGoal

Account level customer lifecycle goal settings.

CustomerManagerLink

Represents customer-manager link relationship.

CustomerNegativeCriterion

A negative criterion for exclusions at the customer level.

CustomerSearchTermInsight

A Customer search term view. Historical data is available starting March 2023.

CustomerSkAdNetworkConversionValueSchema

A CustomerSkAdNetworkConversionValueSchema.

CustomerUserAccess

Represents the permission of a single user onto a single customer.

CustomerUserAccessInvitation

Represent an invitation to a new user on this customer account.

CustomizerAttribute

A customizer attribute. Use CustomerCustomizer, CampaignCustomizer, AdGroupCustomizer, or AdGroupCriterionCustomizer to associate a customizer attribute and set its value at the customer, campaign, ad group, or ad group criterion level, respectively.

DataPartnerIdentifier

The identifier for Data Partner account.

DataPartnerLinkIdentifier

The identifier for Data Partner account.

DetailPlacementView

A view with metrics aggregated by ad group and URL or YouTube video.

DetailedDemographic

A detailed demographic: a particular interest-based vertical to be targeted to reach users based on long-term life facts.

DisplayKeywordView

A display keyword view.

DistanceView

A distance view with metrics aggregated by the user’s distance from an advertiser’s location extensions. Each DistanceBucket includes all impressions that fall within its distance and a single impression will contribute to the metrics for all DistanceBuckets that include the user’s distance.

DomainCategory

A category generated automatically by crawling a domain. If a campaign uses the DynamicSearchAdsSetting, then domain categories will be generated for the domain. The categories can be targeted using WebpageConditionInfo. See: https://support.google.com/google-ads/answer/2471185

DynamicSearchAdsSearchTermView

A dynamic search ads search term view.

ExpandedLandingPageView

A landing page view with metrics aggregated at the expanded final URL level.

Experiment

A Google ads experiment for users to experiment changes on multiple campaigns, compare the performance, and apply the effective changes.

ExperimentArm

A Google ads experiment for users to experiment changes on multiple campaigns, compare the performance, and apply the effective changes.

ExtensionFeedItem

An extension feed item.

Feed

A feed.

FeedAttribute

FeedAttributes define the types of data expected to be present in a Feed. A single FeedAttribute specifies the expected type of the FeedItemAttributes with the same FeedAttributeId. Optionally, a FeedAttribute can be marked as being part of a FeedItem’s unique key.

FeedAttributeOperation

Operation to be performed on a feed attribute list in a mutate.

FeedItem

A feed item.

FeedItemAttributeValue

A feed item attribute value.

FeedItemPlaceholderPolicyInfo

Policy, validation, and quality approval info for a feed item for the specified placeholder type.

FeedItemSet

Represents a set of feed items. The set can be used and shared among certain feed item features. For instance, the set can be referenced within the matching functions of CustomerFeed, CampaignFeed, and AdGroupFeed.

FeedItemSetLink

Represents a link between a FeedItem and a FeedItemSet.

FeedItemTarget

A feed item target.

FeedItemValidationError

Stores a validation error and the set of offending feed attributes which together are responsible for causing a feed item validation error.

FeedMapping

A feed mapping.

FeedPlaceholderView

A feed placeholder view.

GenderView

A gender view.

GeoTargetConstant

A geo target constant.

GeographicView

A geographic view.

GoogleAdsField

A field or resource (artifact) used by GoogleAdsService.

GoogleAdsIdentifier

The identifier for Google Ads account.

GoogleAdsLinkIdentifier

The identifier for Google Ads account.

GranularInsuranceStatus

Insurance status at geo + vertical level.

GranularLicenseStatus

License status at geo + vertical level.

GroupPlacementView

A group placement view.

HotelCenterLinkInvitationIdentifier

The identifier for Hotel account.

HotelGroupView

A hotel group view.

HotelPerformanceView

A hotel performance view.

HotelReconciliation

A hotel reconciliation. It contains conversion information from Hotel bookings to reconcile with advertiser records. These rows may be updated or canceled before billing through Bulk Uploads.

IncomeRangeView

An income range view.

InsuranceVerificationArtifact

A proto holding information specific to a local services insurance.

Invoice

An invoice. All invoice information is snapshotted to match the PDF invoice. For invoices older than the launch of InvoiceService, the snapshotted information may not match the PDF invoice.

KeywordPlan

A Keyword Planner plan. Max number of saved keyword plans: 10000. It’s possible to remove plans if limit is reached.

KeywordPlanAdGroup

A Keyword Planner ad group. Max number of keyword plan ad groups per plan: 200.

KeywordPlanAdGroupKeyword

A Keyword Plan ad group keyword. Max number of keyword plan keywords per plan: 10000.

KeywordPlanCampaign

A Keyword Plan campaign. Max number of keyword plan campaigns per plan allowed: 1.

KeywordPlanCampaignKeyword

A Keyword Plan Campaign keyword. Only negative keywords are supported for Campaign Keyword.

KeywordPlanForecastPeriod

The forecasting period associated with the keyword plan.

KeywordPlanGeoTarget

A geo target.

KeywordThemeConstant

A Smart Campaign keyword theme constant.

KeywordView

A keyword view.

Label

A label.

LandingPageView

A landing page view with metrics aggregated at the unexpanded final URL level.

LanguageConstant

A language.

LeadFormSubmissionData

Data from lead form submissions.

LeadFormSubmissionField

Fields in the submitted lead form.

LicenseVerificationArtifact

A proto holding information specific to a local services license.

LifeEvent

A life event: a particular interest-based vertical to be targeted to reach users when they are in the midst of important life milestones.

ListingGroupFilterDimension

Listing dimensions for the asset group listing group filter.

ListingGroupFilterDimensionPath

The path defining of dimensions defining a listing group filter.

LocalServicesLead

Data from Local Services Lead. Contains details of Lead which is generated when user calls, messages or books service from advertiser. More info: https://ads.google.com/local-services-ads

LocalServicesLeadConversation

Data from Local Services Lead Conversation. Contains details of Lead Conversation which is generated when user calls, messages or books service from advertiser. These are appended to a Lead. More info: https://ads.google.com/local-services-ads

LocalServicesSettings

Settings for Local Services customer.

LocalServicesVerificationArtifact

A local services verification resource.

LocationView

A location view summarizes the performance of campaigns by Location criteria.

ManagedPlacementView

A managed placement view.

MediaAudio

Encapsulates an Audio.

MediaBundle

Represents a ZIP archive media the content of which contains HTML5 assets.

MediaFile

A media file.

MediaImage

Encapsulates an Image.

MediaVideo

Encapsulates a Video.

MerchantCenterIdentifier

The identifier for Google Merchant Center account

MerchantCenterLinkInvitationIdentifier

The identifier for Merchant Center Account.

MessageDetails

Represents details of text message in case of email, message or SMS.

MobileAppCategoryConstant

A mobile application category constant.

MobileDeviceConstant

A mobile device constant.

Note

Represents a note added to a lead by the advertiser. Advertisers can edit notes, which will reset edit time and change description.

OfflineConversionAlert

Alert for offline conversion client summary.

OfflineConversionError

Possible errors for offline conversion client summary.

OfflineConversionSummary

Historical upload summary, grouped by upload date or job.

OfflineConversionUploadClientSummary

Offline conversion upload client summary.

OfflineUserDataJob

A job containing offline user data of store visitors, or user list members that will be processed asynchronously. The uploaded data isn’t readable and the processing results of the job can only be read using GoogleAdsService.Search/SearchStream.

OfflineUserDataJobMetadata

Metadata of offline user data job.

OperatingSystemVersionConstant

A mobile operating system version or a range of versions, depending on operator_type. List of available mobile platforms at https://developers.google.com/google-ads/api/reference/data/codes-formats#mobile-platforms

PaidOrganicSearchTermView

A paid organic search term view providing a view of search stats across ads and organic listings aggregated by search term at the ad group level.

ParentalStatusView

A parental status view.

PaymentsAccount

A payments account, which can be used to set up billing for an Ads customer.

PerStoreView

A per store view. This view provides per store impression reach and local action conversion stats for advertisers.

PhoneCallDetails

Represents details of a phone call conversation.

ProductCategoryConstant

A Product Category.

ProductGroupView

A product group view.

ProductLink

Represents the data sharing connection between a Google Ads customer and another product.

ProductLinkInvitation

Represents an invitation for data sharing connection between a Google Ads account and another account.

QualifyingQuestion

Qualifying Questions for Lead Form.

Recommendation

A recommendation.

RecommendationSubscription

Recommendation Subscription resource

RemarketingAction

A remarketing action. A snippet of JavaScript code that will collect the product id and the type of page people visited (product page, shopping cart page, purchase page, general site visit) on an advertiser’s website.

RemarketingSetting

Remarketing setting for a customer.

SearchTermView

A search term view with metrics aggregated by search term at the ad group level.

SharedCriterion

A criterion belonging to a shared set.

SharedSet

SharedSets are used for sharing criterion exclusions across multiple campaigns.

ShoppingPerformanceView

Shopping performance view. Provides Shopping campaign statistics aggregated at several product dimension levels. Product dimension values from Merchant Center such as brand, category, custom attributes, product condition and product type will reflect the state of each dimension as of the date and time when the corresponding event was recorded.

SmartCampaignSearchTermView

A Smart campaign search term view.

SmartCampaignSetting

Settings for configuring Smart campaigns.

ThirdPartyAppAnalyticsLink

A data sharing connection, allowing the import of third party app analytics into a Google Ads Customer.

ThirdPartyAppAnalyticsLinkIdentifier

The identifiers of a Third Party App Analytics Link.

TopicConstant

Use topics to target or exclude placements in the Google Display Network based on the category into which the placement falls (for example, “Pets & Animals/Pets/Dogs”).

TopicView

A topic view.

TravelActivityGroupView

A travel activity group view.

TravelActivityPerformanceView

A travel activity performance view.

UserInterest

A user interest: a particular interest-based vertical to be targeted.

UserList

A user list. This is a list of users a customer may target.

UserLocationView

A user location view.

Video

A video.

WebpageView

A webpage view.