Enum google_api_proto::google::ads::googleads::v15::resources::ad::AdData

pub enum AdData {

    TextAd(TextAdInfo),
    ExpandedTextAd(ExpandedTextAdInfo),
    CallAd(CallAdInfo),
    ExpandedDynamicSearchAd(ExpandedDynamicSearchAdInfo),
    HotelAd(HotelAdInfo),
    ShoppingSmartAd(ShoppingSmartAdInfo),
    ShoppingProductAd(ShoppingProductAdInfo),
    ImageAd(ImageAdInfo),
    VideoAd(VideoAdInfo),
    VideoResponsiveAd(VideoResponsiveAdInfo),
    ResponsiveSearchAd(ResponsiveSearchAdInfo),
    LegacyResponsiveDisplayAd(LegacyResponsiveDisplayAdInfo),
    AppAd(AppAdInfo),
    LegacyAppInstallAd(LegacyAppInstallAdInfo),
    ResponsiveDisplayAd(ResponsiveDisplayAdInfo),
    LocalAd(LocalAdInfo),
    DisplayUploadAd(DisplayUploadAdInfo),
    AppEngagementAd(AppEngagementAdInfo),
    ShoppingComparisonListingAd(ShoppingComparisonListingAdInfo),
    SmartCampaignAd(SmartCampaignAdInfo),
    AppPreRegistrationAd(AppPreRegistrationAdInfo),
    DiscoveryMultiAssetAd(DiscoveryMultiAssetAdInfo),
    DiscoveryCarouselAd(DiscoveryCarouselAdInfo),
    DiscoveryVideoResponsiveAd(DiscoveryVideoResponsiveAdInfo),
    TravelAd(TravelAdInfo),

}

Details pertinent to the ad type. Exactly one value must be set.

Variants

TextAd(TextAdInfo)

Immutable. Details pertaining to a text ad.

ExpandedTextAd(ExpandedTextAdInfo)

Details pertaining to an expanded text ad.

CallAd(CallAdInfo)

Details pertaining to a call ad.

ExpandedDynamicSearchAd(ExpandedDynamicSearchAdInfo)

Immutable. Details pertaining to an Expanded Dynamic Search Ad. This type of ad has its headline, final URLs, and display URL auto-generated at serving time according to domain name specific information provided by dynamic_search_ads_setting linked at the campaign level.

HotelAd(HotelAdInfo)

Details pertaining to a hotel ad.

ShoppingSmartAd(ShoppingSmartAdInfo)

Details pertaining to a Smart Shopping ad.

ShoppingProductAd(ShoppingProductAdInfo)

Details pertaining to a Shopping product ad.

ImageAd(ImageAdInfo)

Immutable. Details pertaining to an Image ad.

VideoAd(VideoAdInfo)

Details pertaining to a Video ad.

VideoResponsiveAd(VideoResponsiveAdInfo)

Details pertaining to a Video responsive ad.

ResponsiveSearchAd(ResponsiveSearchAdInfo)

Details pertaining to a responsive search ad.

LegacyResponsiveDisplayAd(LegacyResponsiveDisplayAdInfo)

Details pertaining to a legacy responsive display ad.

AppAd(AppAdInfo)

Details pertaining to an app ad.

LegacyAppInstallAd(LegacyAppInstallAdInfo)

Immutable. Details pertaining to a legacy app install ad.

ResponsiveDisplayAd(ResponsiveDisplayAdInfo)

Details pertaining to a responsive display ad.

LocalAd(LocalAdInfo)

Details pertaining to a local ad.

DisplayUploadAd(DisplayUploadAdInfo)

Details pertaining to a display upload ad.

AppEngagementAd(AppEngagementAdInfo)

Details pertaining to an app engagement ad.

ShoppingComparisonListingAd(ShoppingComparisonListingAdInfo)

Details pertaining to a Shopping Comparison Listing ad.

SmartCampaignAd(SmartCampaignAdInfo)

Details pertaining to a Smart campaign ad.

AppPreRegistrationAd(AppPreRegistrationAdInfo)

Details pertaining to an app pre-registration ad.

DiscoveryMultiAssetAd(DiscoveryMultiAssetAdInfo)

Details pertaining to a discovery multi asset ad.

DiscoveryCarouselAd(DiscoveryCarouselAdInfo)

Details pertaining to a discovery carousel ad.

DiscoveryVideoResponsiveAd(DiscoveryVideoResponsiveAdInfo)

Details pertaining to a discovery video responsive ad.

TravelAd(TravelAdInfo)

Details pertaining to a travel ad.

Implementations

impl AdData

pub fn encode<B>(&self, buf: &mut B)
where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<AdData>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>
where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.