Enum google_api_proto::google::ads::googleads::v15::resources::recommendation::Recommendation

pub enum Recommendation {

    CampaignBudgetRecommendation(CampaignBudgetRecommendation),
    ForecastingCampaignBudgetRecommendation(CampaignBudgetRecommendation),
    KeywordRecommendation(KeywordRecommendation),
    TextAdRecommendation(TextAdRecommendation),
    TargetCpaOptInRecommendation(TargetCpaOptInRecommendation),
    MaximizeConversionsOptInRecommendation(MaximizeConversionsOptInRecommendation),
    EnhancedCpcOptInRecommendation(EnhancedCpcOptInRecommendation),
    SearchPartnersOptInRecommendation(SearchPartnersOptInRecommendation),
    MaximizeClicksOptInRecommendation(MaximizeClicksOptInRecommendation),
    OptimizeAdRotationRecommendation(OptimizeAdRotationRecommendation),
    KeywordMatchTypeRecommendation(KeywordMatchTypeRecommendation),
    MoveUnusedBudgetRecommendation(MoveUnusedBudgetRecommendation),
    TargetRoasOptInRecommendation(TargetRoasOptInRecommendation),
    ResponsiveSearchAdRecommendation(ResponsiveSearchAdRecommendation),
    MarginalRoiCampaignBudgetRecommendation(CampaignBudgetRecommendation),
    UseBroadMatchKeywordRecommendation(UseBroadMatchKeywordRecommendation),
    ResponsiveSearchAdAssetRecommendation(ResponsiveSearchAdAssetRecommendation),
    UpgradeSmartShoppingCampaignToPerformanceMaxRecommendation(UpgradeSmartShoppingCampaignToPerformanceMaxRecommendation),
    ResponsiveSearchAdImproveAdStrengthRecommendation(ResponsiveSearchAdImproveAdStrengthRecommendation),
    DisplayExpansionOptInRecommendation(DisplayExpansionOptInRecommendation),
    UpgradeLocalCampaignToPerformanceMaxRecommendation(UpgradeLocalCampaignToPerformanceMaxRecommendation),
    RaiseTargetCpaBidTooLowRecommendation(RaiseTargetCpaBidTooLowRecommendation),
    ForecastingSetTargetRoasRecommendation(ForecastingSetTargetRoasRecommendation),
    CalloutAssetRecommendation(CalloutAssetRecommendation),
    SitelinkAssetRecommendation(SitelinkAssetRecommendation),
    CallAssetRecommendation(CallAssetRecommendation),
    ShoppingAddAgeGroupRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddColorRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddGenderRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddGtinRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddMoreIdentifiersRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddSizeRecommendation(ShoppingOfferAttributeRecommendation),
    ShoppingAddProductsToCampaignRecommendation(ShoppingAddProductsToCampaignRecommendation),
    ShoppingFixDisapprovedProductsRecommendation(ShoppingFixDisapprovedProductsRecommendation),
    ShoppingTargetAllOffersRecommendation(ShoppingTargetAllOffersRecommendation),
    ShoppingFixSuspendedMerchantCenterAccountRecommendation(ShoppingMerchantCenterAccountSuspensionRecommendation),
    ShoppingFixMerchantCenterAccountSuspensionWarningRecommendation(ShoppingMerchantCenterAccountSuspensionRecommendation),
    ShoppingMigrateRegularShoppingCampaignOffersToPerformanceMaxRecommendation(ShoppingMigrateRegularShoppingCampaignOffersToPerformanceMaxRecommendation),
    DynamicImageExtensionOptInRecommendation(DynamicImageExtensionOptInRecommendation),
    RaiseTargetCpaRecommendation(RaiseTargetCpaRecommendation),
    LowerTargetRoasRecommendation(LowerTargetRoasRecommendation),
    PerformanceMaxOptInRecommendation(PerformanceMaxOptInRecommendation),
    ImprovePerformanceMaxAdStrengthRecommendation(ImprovePerformanceMaxAdStrengthRecommendation),
    MigrateDynamicSearchAdsCampaignToPerformanceMaxRecommendation(MigrateDynamicSearchAdsCampaignToPerformanceMaxRecommendation),
    ForecastingSetTargetCpaRecommendation(ForecastingSetTargetCpaRecommendation),
    SetTargetCpaRecommendation(ForecastingSetTargetCpaRecommendation),
    SetTargetRoasRecommendation(ForecastingSetTargetRoasRecommendation),

}

The details of recommendation.

Variants

CampaignBudgetRecommendation(CampaignBudgetRecommendation)

Output only. The campaign budget recommendation.

ForecastingCampaignBudgetRecommendation(CampaignBudgetRecommendation)

Output only. The forecasting campaign budget recommendation.

KeywordRecommendation(KeywordRecommendation)

Output only. The keyword recommendation.

TextAdRecommendation(TextAdRecommendation)

Output only. Add expanded text ad recommendation.

TargetCpaOptInRecommendation(TargetCpaOptInRecommendation)

Output only. The TargetCPA opt-in recommendation.

MaximizeConversionsOptInRecommendation(MaximizeConversionsOptInRecommendation)

Output only. The MaximizeConversions Opt-In recommendation.

EnhancedCpcOptInRecommendation(EnhancedCpcOptInRecommendation)

Output only. The Enhanced Cost-Per-Click Opt-In recommendation.

SearchPartnersOptInRecommendation(SearchPartnersOptInRecommendation)

Output only. The Search Partners Opt-In recommendation.

MaximizeClicksOptInRecommendation(MaximizeClicksOptInRecommendation)

Output only. The MaximizeClicks Opt-In recommendation.

OptimizeAdRotationRecommendation(OptimizeAdRotationRecommendation)

Output only. The Optimize Ad Rotation recommendation.

KeywordMatchTypeRecommendation(KeywordMatchTypeRecommendation)

Output only. The keyword match type recommendation.

MoveUnusedBudgetRecommendation(MoveUnusedBudgetRecommendation)

Output only. The move unused budget recommendation.

TargetRoasOptInRecommendation(TargetRoasOptInRecommendation)

Output only. The Target ROAS opt-in recommendation.

ResponsiveSearchAdRecommendation(ResponsiveSearchAdRecommendation)

Output only. The add responsive search ad recommendation.

MarginalRoiCampaignBudgetRecommendation(CampaignBudgetRecommendation)

Output only. The marginal ROI campaign budget recommendation.

UseBroadMatchKeywordRecommendation(UseBroadMatchKeywordRecommendation)

Output only. The use broad match keyword recommendation.

ResponsiveSearchAdAssetRecommendation(ResponsiveSearchAdAssetRecommendation)

Output only. The responsive search ad asset recommendation.

UpgradeSmartShoppingCampaignToPerformanceMaxRecommendation(UpgradeSmartShoppingCampaignToPerformanceMaxRecommendation)

Output only. The upgrade a Smart Shopping campaign to a Performance Max campaign recommendation.

ResponsiveSearchAdImproveAdStrengthRecommendation(ResponsiveSearchAdImproveAdStrengthRecommendation)

Output only. The responsive search ad improve ad strength recommendation.

DisplayExpansionOptInRecommendation(DisplayExpansionOptInRecommendation)

Output only. The Display Expansion opt-in recommendation.

UpgradeLocalCampaignToPerformanceMaxRecommendation(UpgradeLocalCampaignToPerformanceMaxRecommendation)

Output only. The upgrade a Local campaign to a Performance Max campaign recommendation.

RaiseTargetCpaBidTooLowRecommendation(RaiseTargetCpaBidTooLowRecommendation)

Output only. The raise target CPA bid too low recommendation.

ForecastingSetTargetRoasRecommendation(ForecastingSetTargetRoasRecommendation)

Output only. The forecasting set target ROAS recommendation.

CalloutAssetRecommendation(CalloutAssetRecommendation)

Output only. The callout asset recommendation.

SitelinkAssetRecommendation(SitelinkAssetRecommendation)

Output only. The sitelink asset recommendation.

CallAssetRecommendation(CallAssetRecommendation)

Output only. The call asset recommendation.

ShoppingAddAgeGroupRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add age group recommendation.

ShoppingAddColorRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add color recommendation.

ShoppingAddGenderRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add gender recommendation.

ShoppingAddGtinRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add GTIN recommendation.

ShoppingAddMoreIdentifiersRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add more identifiers recommendation.

ShoppingAddSizeRecommendation(ShoppingOfferAttributeRecommendation)

Output only. The shopping add size recommendation.

ShoppingAddProductsToCampaignRecommendation(ShoppingAddProductsToCampaignRecommendation)

Output only. The shopping add products to campaign recommendation.

ShoppingFixDisapprovedProductsRecommendation(ShoppingFixDisapprovedProductsRecommendation)

Output only. The shopping fix disapproved products recommendation.

ShoppingTargetAllOffersRecommendation(ShoppingTargetAllOffersRecommendation)

Output only. The shopping target all offers recommendation.

ShoppingFixSuspendedMerchantCenterAccountRecommendation(ShoppingMerchantCenterAccountSuspensionRecommendation)

Output only. The shopping fix suspended Merchant Center account recommendation.

ShoppingFixMerchantCenterAccountSuspensionWarningRecommendation(ShoppingMerchantCenterAccountSuspensionRecommendation)

Output only. The shopping fix Merchant Center account suspension warning recommendation.

ShoppingMigrateRegularShoppingCampaignOffersToPerformanceMaxRecommendation(ShoppingMigrateRegularShoppingCampaignOffersToPerformanceMaxRecommendation)

Output only. The shopping migrate Regular Shopping Campaign offers to Performance Max recommendation.

DynamicImageExtensionOptInRecommendation(DynamicImageExtensionOptInRecommendation)

Output only. Recommendation to enable dynamic image extensions on the account, allowing Google to find the best images from ad landing pages and complement text ads.

RaiseTargetCpaRecommendation(RaiseTargetCpaRecommendation)

Output only. Recommendation to raise Target CPA.

LowerTargetRoasRecommendation(LowerTargetRoasRecommendation)

Output only. Recommendation to lower Target ROAS.

PerformanceMaxOptInRecommendation(PerformanceMaxOptInRecommendation)

Output only. The Performance Max Opt In recommendation.

ImprovePerformanceMaxAdStrengthRecommendation(ImprovePerformanceMaxAdStrengthRecommendation)

Output only. The improve Performance Max ad strength recommendation.

MigrateDynamicSearchAdsCampaignToPerformanceMaxRecommendation(MigrateDynamicSearchAdsCampaignToPerformanceMaxRecommendation)

Output only. The Dynamic Search Ads to Performance Max migration recommendation.

ForecastingSetTargetCpaRecommendation(ForecastingSetTargetCpaRecommendation)

Output only. The forecasting set target CPA recommendation.

SetTargetCpaRecommendation(ForecastingSetTargetCpaRecommendation)

Output only. The set target CPA recommendation.

SetTargetRoasRecommendation(ForecastingSetTargetRoasRecommendation)

Output only. The set target ROAS recommendation.

Implementations

impl Recommendation

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Recommendation>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.