Enum google_api_proto::google::ads::googleads::v15::services::mutate_operation_response::Response

pub enum Response {

    AdGroupAdLabelResult(MutateAdGroupAdLabelResult),
    AdGroupAdResult(MutateAdGroupAdResult),
    AdGroupAssetResult(MutateAdGroupAssetResult),
    AdGroupBidModifierResult(MutateAdGroupBidModifierResult),
    AdGroupCriterionCustomizerResult(MutateAdGroupCriterionCustomizerResult),
    AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult),
    AdGroupCriterionResult(MutateAdGroupCriterionResult),
    AdGroupCustomizerResult(MutateAdGroupCustomizerResult),
    AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult),
    AdGroupFeedResult(MutateAdGroupFeedResult),
    AdGroupLabelResult(MutateAdGroupLabelResult),
    AdGroupResult(MutateAdGroupResult),
    AdParameterResult(MutateAdParameterResult),
    AdResult(MutateAdResult),
    AssetResult(MutateAssetResult),
    AssetGroupAssetResult(MutateAssetGroupAssetResult),
    AssetGroupListingGroupFilterResult(MutateAssetGroupListingGroupFilterResult),
    AssetGroupSignalResult(MutateAssetGroupSignalResult),
    AssetGroupResult(MutateAssetGroupResult),
    AssetSetAssetResult(MutateAssetSetAssetResult),
    AssetSetResult(MutateAssetSetResult),
    AudienceResult(MutateAudienceResult),
    BiddingDataExclusionResult(MutateBiddingDataExclusionsResult),
    BiddingSeasonalityAdjustmentResult(MutateBiddingSeasonalityAdjustmentsResult),
    BiddingStrategyResult(MutateBiddingStrategyResult),
    CampaignAssetResult(MutateCampaignAssetResult),
    CampaignAssetSetResult(MutateCampaignAssetSetResult),
    CampaignBidModifierResult(MutateCampaignBidModifierResult),
    CampaignBudgetResult(MutateCampaignBudgetResult),
    CampaignConversionGoalResult(MutateCampaignConversionGoalResult),
    CampaignCriterionResult(MutateCampaignCriterionResult),
    CampaignCustomizerResult(MutateCampaignCustomizerResult),
    CampaignDraftResult(MutateCampaignDraftResult),
    CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult),
    CampaignFeedResult(MutateCampaignFeedResult),
    CampaignGroupResult(MutateCampaignGroupResult),
    CampaignLabelResult(MutateCampaignLabelResult),
    CampaignResult(MutateCampaignResult),
    CampaignSharedSetResult(MutateCampaignSharedSetResult),
    ConversionActionResult(MutateConversionActionResult),
    ConversionCustomVariableResult(MutateConversionCustomVariableResult),
    ConversionGoalCampaignConfigResult(MutateConversionGoalCampaignConfigResult),
    ConversionValueRuleResult(MutateConversionValueRuleResult),
    ConversionValueRuleSetResult(MutateConversionValueRuleSetResult),
    CustomConversionGoalResult(MutateCustomConversionGoalResult),
    CustomerAssetResult(MutateCustomerAssetResult),
    CustomerConversionGoalResult(MutateCustomerConversionGoalResult),
    CustomerCustomizerResult(MutateCustomerCustomizerResult),
    CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult),
    CustomerFeedResult(MutateCustomerFeedResult),
    CustomerLabelResult(MutateCustomerLabelResult),
    CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult),
    CustomerResult(MutateCustomerResult),
    CustomizerAttributeResult(MutateCustomizerAttributeResult),
    ExperimentResult(MutateExperimentResult),
    ExperimentArmResult(MutateExperimentArmResult),
    ExtensionFeedItemResult(MutateExtensionFeedItemResult),
    FeedItemResult(MutateFeedItemResult),
    FeedItemSetResult(MutateFeedItemSetResult),
    FeedItemSetLinkResult(MutateFeedItemSetLinkResult),
    FeedItemTargetResult(MutateFeedItemTargetResult),
    FeedMappingResult(MutateFeedMappingResult),
    FeedResult(MutateFeedResult),
    KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult),
    KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult),
    KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult),
    KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult),
    KeywordPlanResult(MutateKeywordPlansResult),
    LabelResult(MutateLabelResult),
    RecommendationSubscriptionResult(MutateRecommendationSubscriptionResult),
    RemarketingActionResult(MutateRemarketingActionResult),
    SharedCriterionResult(MutateSharedCriterionResult),
    SharedSetResult(MutateSharedSetResult),
    SmartCampaignSettingResult(MutateSmartCampaignSettingResult),
    UserListResult(MutateUserListResult),

}

The mutate response.

Variants

AdGroupAdLabelResult(MutateAdGroupAdLabelResult)

The result for the ad group ad label mutate.

AdGroupAdResult(MutateAdGroupAdResult)

The result for the ad group ad mutate.

AdGroupAssetResult(MutateAdGroupAssetResult)

The result for the ad group asset mutate.

AdGroupBidModifierResult(MutateAdGroupBidModifierResult)

The result for the ad group bid modifier mutate.

AdGroupCriterionCustomizerResult(MutateAdGroupCriterionCustomizerResult)

The result for the ad group criterion customizer mutate.

AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult)

The result for the ad group criterion label mutate.

AdGroupCriterionResult(MutateAdGroupCriterionResult)

The result for the ad group criterion mutate.

AdGroupCustomizerResult(MutateAdGroupCustomizerResult)

The result for the ad group customizer mutate.

AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult)

The result for the ad group extension setting mutate.

AdGroupFeedResult(MutateAdGroupFeedResult)

The result for the ad group feed mutate.

AdGroupLabelResult(MutateAdGroupLabelResult)

The result for the ad group label mutate.

AdGroupResult(MutateAdGroupResult)

The result for the ad group mutate.

AdParameterResult(MutateAdParameterResult)

The result for the ad parameter mutate.

AdResult(MutateAdResult)

The result for the ad mutate.

AssetResult(MutateAssetResult)

The result for the asset mutate.

AssetGroupAssetResult(MutateAssetGroupAssetResult)

The result for the asset group asset mutate.

AssetGroupListingGroupFilterResult(MutateAssetGroupListingGroupFilterResult)

The result for the asset group listing group filter mutate.

AssetGroupSignalResult(MutateAssetGroupSignalResult)

The result for the asset group signal mutate.

AssetGroupResult(MutateAssetGroupResult)

The result for the asset group mutate.

AssetSetAssetResult(MutateAssetSetAssetResult)

The result for the asset set asset mutate.

AssetSetResult(MutateAssetSetResult)

The result for the asset set mutate.

AudienceResult(MutateAudienceResult)

The result for the audience mutate.

BiddingDataExclusionResult(MutateBiddingDataExclusionsResult)

The result for the bidding data exclusion mutate.

BiddingSeasonalityAdjustmentResult(MutateBiddingSeasonalityAdjustmentsResult)

The result for the bidding seasonality adjustment mutate.

BiddingStrategyResult(MutateBiddingStrategyResult)

The result for the bidding strategy mutate.

CampaignAssetResult(MutateCampaignAssetResult)

The result for the campaign asset mutate.

CampaignAssetSetResult(MutateCampaignAssetSetResult)

The result for the campaign asset set mutate.

CampaignBidModifierResult(MutateCampaignBidModifierResult)

The result for the campaign bid modifier mutate.

CampaignBudgetResult(MutateCampaignBudgetResult)

The result for the campaign budget mutate.

CampaignConversionGoalResult(MutateCampaignConversionGoalResult)

The result for the campaign conversion goal mutate.

CampaignCriterionResult(MutateCampaignCriterionResult)

The result for the campaign criterion mutate.

CampaignCustomizerResult(MutateCampaignCustomizerResult)

The result for the campaign customizer mutate.

CampaignDraftResult(MutateCampaignDraftResult)

The result for the campaign draft mutate.

CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult)

The result for the campaign extension setting mutate.

CampaignFeedResult(MutateCampaignFeedResult)

The result for the campaign feed mutate.

CampaignGroupResult(MutateCampaignGroupResult)

The result for the campaign group mutate.

CampaignLabelResult(MutateCampaignLabelResult)

The result for the campaign label mutate.

CampaignResult(MutateCampaignResult)

The result for the campaign mutate.

CampaignSharedSetResult(MutateCampaignSharedSetResult)

The result for the campaign shared set mutate.

ConversionActionResult(MutateConversionActionResult)

The result for the conversion action mutate.

ConversionCustomVariableResult(MutateConversionCustomVariableResult)

The result for the conversion custom variable mutate.

ConversionGoalCampaignConfigResult(MutateConversionGoalCampaignConfigResult)

The result for the conversion goal campaign config mutate.

ConversionValueRuleResult(MutateConversionValueRuleResult)

The result for the conversion value rule mutate.

ConversionValueRuleSetResult(MutateConversionValueRuleSetResult)

The result for the conversion value rule set mutate.

CustomConversionGoalResult(MutateCustomConversionGoalResult)

The result for the custom conversion goal mutate.

CustomerAssetResult(MutateCustomerAssetResult)

The result for the customer asset mutate.

CustomerConversionGoalResult(MutateCustomerConversionGoalResult)

The result for the customer conversion goal mutate.

CustomerCustomizerResult(MutateCustomerCustomizerResult)

The result for the customer customizer mutate.

CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult)

The result for the customer extension setting mutate.

CustomerFeedResult(MutateCustomerFeedResult)

The result for the customer feed mutate.

CustomerLabelResult(MutateCustomerLabelResult)

The result for the customer label mutate.

CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult)

The result for the customer negative criterion mutate.

CustomerResult(MutateCustomerResult)

The result for the customer mutate.

CustomizerAttributeResult(MutateCustomizerAttributeResult)

The result for the customizer attribute mutate.

ExperimentResult(MutateExperimentResult)

The result for the experiment mutate.

ExperimentArmResult(MutateExperimentArmResult)

The result for the experiment arm mutate.

ExtensionFeedItemResult(MutateExtensionFeedItemResult)

The result for the extension feed item mutate.

FeedItemResult(MutateFeedItemResult)

The result for the feed item mutate.

FeedItemSetResult(MutateFeedItemSetResult)

The result for the feed item set mutate.

FeedItemSetLinkResult(MutateFeedItemSetLinkResult)

The result for the feed item set link mutate.

FeedItemTargetResult(MutateFeedItemTargetResult)

The result for the feed item target mutate.

FeedMappingResult(MutateFeedMappingResult)

The result for the feed mapping mutate.

FeedResult(MutateFeedResult)

The result for the feed mutate.

KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult)

The result for the keyword plan ad group mutate.

KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult)

The result for the keyword plan campaign mutate.

KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult)

The result for the keyword plan ad group keyword mutate.

KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult)

The result for the keyword plan campaign keyword mutate.

KeywordPlanResult(MutateKeywordPlansResult)

The result for the keyword plan mutate.

LabelResult(MutateLabelResult)

The result for the label mutate.

RecommendationSubscriptionResult(MutateRecommendationSubscriptionResult)

The result for the recommendation subscription mutate.

RemarketingActionResult(MutateRemarketingActionResult)

The result for the remarketing action mutate.

SharedCriterionResult(MutateSharedCriterionResult)

The result for the shared criterion mutate.

SharedSetResult(MutateSharedSetResult)

The result for the shared set mutate.

SmartCampaignSettingResult(MutateSmartCampaignSettingResult)

The result for the Smart campaign setting mutate.

UserListResult(MutateUserListResult)

The result for the user list mutate.

Implementations

impl Response

pub fn encode<B>(&self, buf: &mut B)
where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<Response>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>
where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.