Struct google_api_proto::google::ads::googleads::v15::services::GenerateKeywordIdeasRequest

pub struct GenerateKeywordIdeasRequest {
    pub customer_id: String,
    pub language: Option<String>,
    pub geo_target_constants: Vec<String>,
    pub include_adult_keywords: bool,
    pub page_token: String,
    pub page_size: i32,
    pub keyword_plan_network: i32,
    pub keyword_annotation: Vec<i32>,
    pub aggregate_metrics: Option<KeywordPlanAggregateMetrics>,
    pub historical_metrics_options: Option<HistoricalMetricsOptions>,
    pub seed: Option<Seed>,
}

Request message for [KeywordPlanIdeaService.GenerateKeywordIdeas][google.ads.googleads.v15.services.KeywordPlanIdeaService.GenerateKeywordIdeas].

Fields
customer_id: String

The ID of the customer with the recommendation.

language: Option<String>

The resource name of the language to target. Each keyword belongs to some set of languages; a keyword is included if language is one of its languages. If not set, all keywords will be included.

geo_target_constants: Vec<String>

The resource names of the location to target. Maximum is 10. An empty list MAY be used to specify all targeting geos.

include_adult_keywords: bool

If true, adult keywords will be included in response. The default value is false.

page_token: String

Token of the page to retrieve. If not specified, the first page of results will be returned. To request next page of results use the value obtained from next_page_token in the previous response. The request fields must match across pages.

page_size: i32

Number of results to retrieve in a single page. A maximum of 10,000 results may be returned, if the page_size exceeds this, it is ignored. If unspecified, at most 10,000 results will be returned. The server may decide to further limit the number of returned resources. If the response contains fewer than 10,000 results it may not be assumed as last page of results.

keyword_plan_network: i32

Targeting network. If not set, Google Search And Partners Network will be used.

keyword_annotation: Vec<i32>

The keyword annotations to include in response.

aggregate_metrics: Option<KeywordPlanAggregateMetrics>

The aggregate fields to include in response.

historical_metrics_options: Option<HistoricalMetricsOptions>

The options for historical metrics data.

seed: Option<Seed>

The type of seed to generate keyword ideas.

Implementations

impl GenerateKeywordIdeasRequest

pub fn keyword_plan_network(&self) -> KeywordPlanNetwork

Returns the enum value of keyword_plan_network, or the default if the field is set to an invalid enum value.

pub fn set_keyword_plan_network(&mut self, value: KeywordPlanNetwork)

Sets keyword_plan_network to the provided enum value.

pub fn language(&self) -> &str

Returns the value of language, or the default value if language is unset.

pub fn keyword_annotation( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<KeywordPlanKeywordAnnotation>>

Returns an iterator which yields the valid enum values contained in keyword_annotation.

pub fn push_keyword_annotation(&mut self, value: KeywordPlanKeywordAnnotation)

Appends the provided enum value to keyword_annotation.