Enum google_api_proto::google::ads::googleads::v16::enums::conversion_or_adjustment_lag_bucket_enum::ConversionOrAdjustmentLagBucket

#[repr(i32)]
pub enum ConversionOrAdjustmentLagBucket {

    Unspecified = 0,
    Unknown = 1,
    ConversionLessThanOneDay = 2,
    ConversionOneToTwoDays = 3,
    ConversionTwoToThreeDays = 4,
    ConversionThreeToFourDays = 5,
    ConversionFourToFiveDays = 6,
    ConversionFiveToSixDays = 7,
    ConversionSixToSevenDays = 8,
    ConversionSevenToEightDays = 9,
    ConversionEightToNineDays = 10,
    ConversionNineToTenDays = 11,
    ConversionTenToElevenDays = 12,
    ConversionElevenToTwelveDays = 13,
    ConversionTwelveToThirteenDays = 14,
    ConversionThirteenToFourteenDays = 15,
    ConversionFourteenToTwentyOneDays = 16,
    ConversionTwentyOneToThirtyDays = 17,
    ConversionThirtyToFortyFiveDays = 18,
    ConversionFortyFiveToSixtyDays = 19,
    ConversionSixtyToNinetyDays = 20,
    AdjustmentLessThanOneDay = 21,
    AdjustmentOneToTwoDays = 22,
    AdjustmentTwoToThreeDays = 23,
    AdjustmentThreeToFourDays = 24,
    AdjustmentFourToFiveDays = 25,
    AdjustmentFiveToSixDays = 26,
    AdjustmentSixToSevenDays = 27,
    AdjustmentSevenToEightDays = 28,
    AdjustmentEightToNineDays = 29,
    AdjustmentNineToTenDays = 30,
    AdjustmentTenToElevenDays = 31,
    AdjustmentElevenToTwelveDays = 32,
    AdjustmentTwelveToThirteenDays = 33,
    AdjustmentThirteenToFourteenDays = 34,
    AdjustmentFourteenToTwentyOneDays = 35,
    AdjustmentTwentyOneToThirtyDays = 36,
    AdjustmentThirtyToFortyFiveDays = 37,
    AdjustmentFortyFiveToSixtyDays = 38,
    AdjustmentSixtyToNinetyDays = 39,
    AdjustmentNinetyToOneHundredAndFortyFiveDays = 40,
    ConversionUnknown = 41,
    AdjustmentUnknown = 42,

}

Enum representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

ConversionLessThanOneDay = 2

Conversion lag bucket from 0 to 1 day. 0 day is included, 1 day is not.

ConversionOneToTwoDays = 3

Conversion lag bucket from 1 to 2 days. 1 day is included, 2 days is not.

ConversionTwoToThreeDays = 4

Conversion lag bucket from 2 to 3 days. 2 days is included, 3 days is not.

ConversionThreeToFourDays = 5

Conversion lag bucket from 3 to 4 days. 3 days is included, 4 days is not.

ConversionFourToFiveDays = 6

Conversion lag bucket from 4 to 5 days. 4 days is included, 5 days is not.

ConversionFiveToSixDays = 7

Conversion lag bucket from 5 to 6 days. 5 days is included, 6 days is not.

ConversionSixToSevenDays = 8

Conversion lag bucket from 6 to 7 days. 6 days is included, 7 days is not.

ConversionSevenToEightDays = 9

Conversion lag bucket from 7 to 8 days. 7 days is included, 8 days is not.

ConversionEightToNineDays = 10

Conversion lag bucket from 8 to 9 days. 8 days is included, 9 days is not.

ConversionNineToTenDays = 11

Conversion lag bucket from 9 to 10 days. 9 days is included, 10 days is not.

ConversionTenToElevenDays = 12

Conversion lag bucket from 10 to 11 days. 10 days is included, 11 days is not.

ConversionElevenToTwelveDays = 13

Conversion lag bucket from 11 to 12 days. 11 days is included, 12 days is not.

ConversionTwelveToThirteenDays = 14

Conversion lag bucket from 12 to 13 days. 12 days is included, 13 days is not.

ConversionThirteenToFourteenDays = 15

Conversion lag bucket from 13 to 14 days. 13 days is included, 14 days is not.

ConversionFourteenToTwentyOneDays = 16

Conversion lag bucket from 14 to 21 days. 14 days is included, 21 days is not.

ConversionTwentyOneToThirtyDays = 17

Conversion lag bucket from 21 to 30 days. 21 days is included, 30 days is not.

ConversionThirtyToFortyFiveDays = 18

Conversion lag bucket from 30 to 45 days. 30 days is included, 45 days is not.

ConversionFortyFiveToSixtyDays = 19

Conversion lag bucket from 45 to 60 days. 45 days is included, 60 days is not.

ConversionSixtyToNinetyDays = 20

Conversion lag bucket from 60 to 90 days. 60 days is included, 90 days is not.

AdjustmentLessThanOneDay = 21

Conversion adjustment lag bucket from 0 to 1 day. 0 day is included, 1 day is not.

AdjustmentOneToTwoDays = 22

Conversion adjustment lag bucket from 1 to 2 days. 1 day is included, 2 days is not.

AdjustmentTwoToThreeDays = 23

Conversion adjustment lag bucket from 2 to 3 days. 2 days is included, 3 days is not.

AdjustmentThreeToFourDays = 24

Conversion adjustment lag bucket from 3 to 4 days. 3 days is included, 4 days is not.

AdjustmentFourToFiveDays = 25

Conversion adjustment lag bucket from 4 to 5 days. 4 days is included, 5 days is not.

AdjustmentFiveToSixDays = 26

Conversion adjustment lag bucket from 5 to 6 days. 5 days is included, 6 days is not.

AdjustmentSixToSevenDays = 27

Conversion adjustment lag bucket from 6 to 7 days. 6 days is included, 7 days is not.

AdjustmentSevenToEightDays = 28

Conversion adjustment lag bucket from 7 to 8 days. 7 days is included, 8 days is not.

AdjustmentEightToNineDays = 29

Conversion adjustment lag bucket from 8 to 9 days. 8 days is included, 9 days is not.

AdjustmentNineToTenDays = 30

Conversion adjustment lag bucket from 9 to 10 days. 9 days is included, 10 days is not.

AdjustmentTenToElevenDays = 31

Conversion adjustment lag bucket from 10 to 11 days. 10 days is included, 11 days is not.

AdjustmentElevenToTwelveDays = 32

Conversion adjustment lag bucket from 11 to 12 days. 11 days is included, 12 days is not.

AdjustmentTwelveToThirteenDays = 33

Conversion adjustment lag bucket from 12 to 13 days. 12 days is included, 13 days is not.

AdjustmentThirteenToFourteenDays = 34

Conversion adjustment lag bucket from 13 to 14 days. 13 days is included, 14 days is not.

AdjustmentFourteenToTwentyOneDays = 35

Conversion adjustment lag bucket from 14 to 21 days. 14 days is included, 21 days is not.

AdjustmentTwentyOneToThirtyDays = 36

Conversion adjustment lag bucket from 21 to 30 days. 21 days is included, 30 days is not.

AdjustmentThirtyToFortyFiveDays = 37

Conversion adjustment lag bucket from 30 to 45 days. 30 days is included, 45 days is not.

AdjustmentFortyFiveToSixtyDays = 38

Conversion adjustment lag bucket from 45 to 60 days. 45 days is included, 60 days is not.

AdjustmentSixtyToNinetyDays = 39

Conversion adjustment lag bucket from 60 to 90 days. 60 days is included, 90 days is not.

AdjustmentNinetyToOneHundredAndFortyFiveDays = 40

Conversion adjustment lag bucket from 90 to 145 days. 90 days is included, 145 days is not.

ConversionUnknown = 41

Conversion lag bucket UNKNOWN. This is for dates before conversion lag bucket was available in Google Ads.

AdjustmentUnknown = 42

Conversion adjustment lag bucket UNKNOWN. This is for dates before conversion adjustment lag bucket was available in Google Ads.

Implementations

impl ConversionOrAdjustmentLagBucket

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionOrAdjustmentLagBucket.

pub fn from_i32(value: i32) -> Option<ConversionOrAdjustmentLagBucket>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ConversionOrAdjustmentLagBucket, or None if value is not a valid variant.

impl ConversionOrAdjustmentLagBucket

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.