Enum google_api_proto::google::ads::googleads::v16::enums::lead_form_field_user_input_type_enum::LeadFormFieldUserInputType

#[repr(i32)]
pub enum LeadFormFieldUserInputType {

    Unspecified = 0,
    Unknown = 1,
    FullName = 2,
    Email = 3,
    PhoneNumber = 4,
    PostalCode = 5,
    StreetAddress = 8,
    City = 9,
    Region = 10,
    Country = 11,
    WorkEmail = 12,
    CompanyName = 13,
    WorkPhone = 14,
    JobTitle = 15,
    GovernmentIssuedIdCpfBr = 16,
    GovernmentIssuedIdDniAr = 17,
    GovernmentIssuedIdDniPe = 18,
    GovernmentIssuedIdRutCl = 19,
    GovernmentIssuedIdCcCo = 20,
    GovernmentIssuedIdCiEc = 21,
    GovernmentIssuedIdRfcMx = 22,
    FirstName = 23,
    LastName = 24,
    VehicleModel = 1_001,
    VehicleType = 1_002,
    PreferredDealership = 1_003,
    VehiclePurchaseTimeline = 1_004,
    VehicleOwnership = 1_005,
    VehiclePaymentType = 1_009,
    VehicleCondition = 1_010,
    CompanySize = 1_006,
    AnnualSales = 1_007,
    YearsInBusiness = 1_008,
    JobDepartment = 1_011,
    JobRole = 1_012,
    Over18Age = 1_078,
    Over19Age = 1_079,
    Over20Age = 1_080,
    Over21Age = 1_081,
    Over22Age = 1_082,
    Over23Age = 1_083,
    Over24Age = 1_084,
    Over25Age = 1_085,
    Over26Age = 1_086,
    Over27Age = 1_087,
    Over28Age = 1_088,
    Over29Age = 1_089,
    Over30Age = 1_090,
    Over31Age = 1_091,
    Over32Age = 1_092,
    Over33Age = 1_093,
    Over34Age = 1_094,
    Over35Age = 1_095,
    Over36Age = 1_096,
    Over37Age = 1_097,
    Over38Age = 1_098,
    Over39Age = 1_099,
    Over40Age = 1_100,
    Over41Age = 1_101,
    Over42Age = 1_102,
    Over43Age = 1_103,
    Over44Age = 1_104,
    Over45Age = 1_105,
    Over46Age = 1_106,
    Over47Age = 1_107,
    Over48Age = 1_108,
    Over49Age = 1_109,
    Over50Age = 1_110,
    Over51Age = 1_111,
    Over52Age = 1_112,
    Over53Age = 1_113,
    Over54Age = 1_114,
    Over55Age = 1_115,
    Over56Age = 1_116,
    Over57Age = 1_117,
    Over58Age = 1_118,
    Over59Age = 1_119,
    Over60Age = 1_120,
    Over61Age = 1_121,
    Over62Age = 1_122,
    Over63Age = 1_123,
    Over64Age = 1_124,
    Over65Age = 1_125,
    EducationProgram = 1_013,
    EducationCourse = 1_014,
    Product = 1_016,
    Service = 1_017,
    Offer = 1_018,
    Category = 1_019,
    PreferredContactMethod = 1_020,
    PreferredLocation = 1_021,
    PreferredContactTime = 1_022,
    PurchaseTimeline = 1_023,
    YearsOfExperience = 1_048,
    JobIndustry = 1_049,
    LevelOfEducation = 1_050,
    PropertyType = 1_024,
    RealtorHelpGoal = 1_025,
    PropertyCommunity = 1_026,
    PriceRange = 1_027,
    NumberOfBedrooms = 1_028,
    FurnishedProperty = 1_029,
    PetsAllowedProperty = 1_030,
    NextPlannedPurchase = 1_031,
    EventSignupInterest = 1_033,
    PreferredShoppingPlaces = 1_034,
    FavoriteBrand = 1_035,
    TransportationCommercialLicenseType = 1_036,
    EventBookingInterest = 1_038,
    DestinationCountry = 1_039,
    DestinationCity = 1_040,
    DepartureCountry = 1_041,
    DepartureCity = 1_042,
    DepartureDate = 1_043,
    ReturnDate = 1_044,
    NumberOfTravelers = 1_045,
    TravelBudget = 1_046,
    TravelAccommodation = 1_047,

}

Enum describing the input type of a lead form field.

Variants

Unspecified = 0

Not specified.

Unknown = 1

Used for return value only. Represents value unknown in this version.

FullName = 2

The user will be asked to fill in their given and family name. This field cannot be set at the same time as GIVEN_NAME or FAMILY_NAME.

Email = 3

The user will be asked to fill in their email address.

PhoneNumber = 4

The user will be asked to fill in their phone number.

PostalCode = 5

The user will be asked to fill in their zip code.

StreetAddress = 8

The user will be asked to fill in their street address.

City = 9

The user will be asked to fill in their city.

Region = 10

The user will be asked to fill in their region part of the address (for example, state for US, province for Canada).

Country = 11

The user will be asked to fill in their country.

WorkEmail = 12

The user will be asked to fill in their work email address.

CompanyName = 13

The user will be asked to fill in their company name.

WorkPhone = 14

The user will be asked to fill in their work phone.

JobTitle = 15

The user will be asked to fill in their job title.

GovernmentIssuedIdCpfBr = 16

The user will be asked to fill in their CPF for Brazil users.

GovernmentIssuedIdDniAr = 17

The user will be asked to fill in their DNI for Argentina users.

GovernmentIssuedIdDniPe = 18

The user will be asked to fill in their DNI for Peru users.

GovernmentIssuedIdRutCl = 19

The user will be asked to fill in their RUT for Chile users.

GovernmentIssuedIdCcCo = 20

The user will be asked to fill in their CC for Colombia users.

GovernmentIssuedIdCiEc = 21

The user will be asked to fill in their CI for Ecuador users.

GovernmentIssuedIdRfcMx = 22

The user will be asked to fill in their RFC for Mexico users.

FirstName = 23

The user will be asked to fill in their first name. This field can not be set at the same time as FULL_NAME.

LastName = 24

The user will be asked to fill in their last name. This field can not be set at the same time as FULL_NAME.

VehicleModel = 1_001

Question: “Which model are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleType = 1_002

Question: “Which type of vehicle are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredDealership = 1_003

Question: “What is your preferred dealership?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehiclePurchaseTimeline = 1_004

Question: “When do you plan on purchasing a vehicle?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleOwnership = 1_005

Question: “Do you own a vehicle?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehiclePaymentType = 1_009

Question: “What vehicle ownership option are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

VehicleCondition = 1_010

Question: “What type of vehicle condition are you interested in?” Category: “Auto” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

CompanySize = 1_006

Question: “What size is your company?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

AnnualSales = 1_007

Question: “What is your annual sales volume?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

YearsInBusiness = 1_008

Question: “How many years have you been in business?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobDepartment = 1_011

Question: “What is your job department?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobRole = 1_012

Question: “What is your job role?” Category: “Business” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over18Age = 1_078

Question: “Are you over 18 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over19Age = 1_079

Question: “Are you over 19 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over20Age = 1_080

Question: “Are you over 20 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over21Age = 1_081

Question: “Are you over 21 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over22Age = 1_082

Question: “Are you over 22 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over23Age = 1_083

Question: “Are you over 23 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over24Age = 1_084

Question: “Are you over 24 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over25Age = 1_085

Question: “Are you over 25 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over26Age = 1_086

Question: “Are you over 26 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over27Age = 1_087

Question: “Are you over 27 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over28Age = 1_088

Question: “Are you over 28 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over29Age = 1_089

Question: “Are you over 29 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over30Age = 1_090

Question: “Are you over 30 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over31Age = 1_091

Question: “Are you over 31 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over32Age = 1_092

Question: “Are you over 32 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over33Age = 1_093

Question: “Are you over 33 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over34Age = 1_094

Question: “Are you over 34 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over35Age = 1_095

Question: “Are you over 35 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over36Age = 1_096

Question: “Are you over 36 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over37Age = 1_097

Question: “Are you over 37 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over38Age = 1_098

Question: “Are you over 38 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over39Age = 1_099

Question: “Are you over 39 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over40Age = 1_100

Question: “Are you over 40 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over41Age = 1_101

Question: “Are you over 41 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over42Age = 1_102

Question: “Are you over 42 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over43Age = 1_103

Question: “Are you over 43 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over44Age = 1_104

Question: “Are you over 44 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over45Age = 1_105

Question: “Are you over 45 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over46Age = 1_106

Question: “Are you over 46 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over47Age = 1_107

Question: “Are you over 47 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over48Age = 1_108

Question: “Are you over 48 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over49Age = 1_109

Question: “Are you over 49 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over50Age = 1_110

Question: “Are you over 50 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over51Age = 1_111

Question: “Are you over 51 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over52Age = 1_112

Question: “Are you over 52 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over53Age = 1_113

Question: “Are you over 53 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over54Age = 1_114

Question: “Are you over 54 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over55Age = 1_115

Question: “Are you over 55 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over56Age = 1_116

Question: “Are you over 56 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over57Age = 1_117

Question: “Are you over 57 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over58Age = 1_118

Question: “Are you over 58 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over59Age = 1_119

Question: “Are you over 59 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over60Age = 1_120

Question: “Are you over 60 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over61Age = 1_121

Question: “Are you over 61 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over62Age = 1_122

Question: “Are you over 62 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over63Age = 1_123

Question: “Are you over 63 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over64Age = 1_124

Question: “Are you over 64 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Over65Age = 1_125

Question: “Are you over 65 years of age?” Category: “Demographics” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EducationProgram = 1_013

Question: “Which program are you interested in?” Category: “Education” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EducationCourse = 1_014

Question: “Which course are you interested in?” Category: “Education” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Product = 1_016

Question: “Which product are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Service = 1_017

Question: “Which service are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Offer = 1_018

Question: “Which offer are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Category = 1_019

Question: “Which category are you interested in?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredContactMethod = 1_020

Question: “What is your preferred method of contact?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredLocation = 1_021

Question: “What is your preferred location?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredContactTime = 1_022

Question: “What is the best time to contact you?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PurchaseTimeline = 1_023

Question: “When are you looking to make a purchase?” Category: “General” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

YearsOfExperience = 1_048

Question: “How many years of work experience do you have?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

JobIndustry = 1_049

Question: “What industry do you work in?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

LevelOfEducation = 1_050

Question: “What is your highest level of education?” Category: “Jobs” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PropertyType = 1_024

Question: “What type of property are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

RealtorHelpGoal = 1_025

Question: “What do you need a realtor’s help with?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PropertyCommunity = 1_026

Question: “What neighborhood are you interested in?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PriceRange = 1_027

Question: “What price range are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NumberOfBedrooms = 1_028

Question: “How many bedrooms are you looking for?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

FurnishedProperty = 1_029

Question: “Are you looking for a fully furnished property?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PetsAllowedProperty = 1_030

Question: “Are you looking for properties that allow pets?” Category: “Real Estate” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NextPlannedPurchase = 1_031

Question: “What is the next product you plan to purchase?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EventSignupInterest = 1_033

Question: “Would you like to sign up for an event?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

PreferredShoppingPlaces = 1_034

Question: “Where are you interested in shopping?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

FavoriteBrand = 1_035

Question: “What is your favorite brand?” Category: “Retail” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TransportationCommercialLicenseType = 1_036

Question: “Which type of valid commercial license do you have?” Category: “Transportation” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

EventBookingInterest = 1_038

Question: “Interested in booking an event?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DestinationCountry = 1_039

Question: “What is your destination country?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DestinationCity = 1_040

Question: “What is your destination city?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureCountry = 1_041

Question: “What is your departure country?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureCity = 1_042

Question: “What is your departure city?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

DepartureDate = 1_043

Question: “What is your departure date?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

ReturnDate = 1_044

Question: “What is your return date?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

NumberOfTravelers = 1_045

Question: “How many people are you traveling with?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TravelBudget = 1_046

Question: “What is your travel budget?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

TravelAccommodation = 1_047

Question: “Where do you want to stay during your travel?” Category: “Travel” This field is subject to a limit of 5 qualifying questions per form and cannot be used if values are set using custom_question_fields.

Implementations

impl LeadFormFieldUserInputType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of LeadFormFieldUserInputType.

pub fn from_i32(value: i32) -> Option<LeadFormFieldUserInputType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a LeadFormFieldUserInputType, or None if value is not a valid variant.

impl LeadFormFieldUserInputType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.