Enum google_api_proto::google::ads::googleads::v16::enums::resource_limit_type_enum::ResourceLimitType

#[repr(i32)]
pub enum ResourceLimitType {

    Unspecified = 0,
    Unknown = 1,
    CampaignsPerCustomer = 2,
    BaseCampaignsPerCustomer = 3,
    ExperimentCampaignsPerCustomer = 105,
    HotelCampaignsPerCustomer = 4,
    SmartShoppingCampaignsPerCustomer = 5,
    AdGroupsPerCampaign = 6,
    AdGroupsPerShoppingCampaign = 8,
    AdGroupsPerHotelCampaign = 9,
    ReportingAdGroupsPerLocalCampaign = 10,
    ReportingAdGroupsPerAppCampaign = 11,
    ManagedAdGroupsPerSmartCampaign = 52,
    AdGroupCriteriaPerCustomer = 12,
    BaseAdGroupCriteriaPerCustomer = 13,
    ExperimentAdGroupCriteriaPerCustomer = 107,
    AdGroupCriteriaPerCampaign = 14,
    CampaignCriteriaPerCustomer = 15,
    BaseCampaignCriteriaPerCustomer = 16,
    ExperimentCampaignCriteriaPerCustomer = 108,
    WebpageCriteriaPerCustomer = 17,
    BaseWebpageCriteriaPerCustomer = 18,
    ExperimentWebpageCriteriaPerCustomer = 19,
    CombinedAudienceCriteriaPerAdGroup = 20,
    CustomerNegativePlacementCriteriaPerCustomer = 21,
    CustomerNegativeYoutubeChannelCriteriaPerCustomer = 22,
    CriteriaPerAdGroup = 23,
    ListingGroupsPerAdGroup = 24,
    ExplicitlySharedBudgetsPerCustomer = 25,
    ImplicitlySharedBudgetsPerCustomer = 26,
    CombinedAudienceCriteriaPerCampaign = 27,
    NegativeKeywordsPerCampaign = 28,
    NegativePlacementsPerCampaign = 29,
    GeoTargetsPerCampaign = 30,
    NegativeIpBlocksPerCampaign = 32,
    ProximitiesPerCampaign = 33,
    ListingScopesPerShoppingCampaign = 34,
    ListingScopesPerNonShoppingCampaign = 35,
    NegativeKeywordsPerSharedSet = 36,
    NegativePlacementsPerSharedSet = 37,
    SharedSetsPerCustomerForTypeDefault = 40,
    SharedSetsPerCustomerForNegativePlacementListLower = 41,
    HotelAdvanceBookingWindowBidModifiersPerAdGroup = 44,
    BiddingStrategiesPerCustomer = 45,
    BasicUserListsPerCustomer = 47,
    LogicalUserListsPerCustomer = 48,
    RuleBasedUserListsPerCustomer = 153,
    BaseAdGroupAdsPerCustomer = 53,
    ExperimentAdGroupAdsPerCustomer = 54,
    AdGroupAdsPerCampaign = 55,
    TextAndOtherAdsPerAdGroup = 56,
    ImageAdsPerAdGroup = 57,
    ShoppingSmartAdsPerAdGroup = 58,
    ResponsiveSearchAdsPerAdGroup = 59,
    AppAdsPerAdGroup = 60,
    AppEngagementAdsPerAdGroup = 61,
    LocalAdsPerAdGroup = 62,
    VideoAdsPerAdGroup = 63,
    LeadFormCampaignAssetsPerCampaign = 143,
    PromotionCustomerAssetsPerCustomer = 79,
    PromotionCampaignAssetsPerCampaign = 80,
    PromotionAdGroupAssetsPerAdGroup = 81,
    CalloutCustomerAssetsPerCustomer = 134,
    CalloutCampaignAssetsPerCampaign = 135,
    CalloutAdGroupAssetsPerAdGroup = 136,
    SitelinkCustomerAssetsPerCustomer = 137,
    SitelinkCampaignAssetsPerCampaign = 138,
    SitelinkAdGroupAssetsPerAdGroup = 139,
    StructuredSnippetCustomerAssetsPerCustomer = 140,
    StructuredSnippetCampaignAssetsPerCampaign = 141,
    StructuredSnippetAdGroupAssetsPerAdGroup = 142,
    MobileAppCustomerAssetsPerCustomer = 144,
    MobileAppCampaignAssetsPerCampaign = 145,
    MobileAppAdGroupAssetsPerAdGroup = 146,
    HotelCalloutCustomerAssetsPerCustomer = 147,
    HotelCalloutCampaignAssetsPerCampaign = 148,
    HotelCalloutAdGroupAssetsPerAdGroup = 149,
    CallCustomerAssetsPerCustomer = 150,
    CallCampaignAssetsPerCampaign = 151,
    CallAdGroupAssetsPerAdGroup = 152,
    PriceCustomerAssetsPerCustomer = 154,
    PriceCampaignAssetsPerCampaign = 155,
    PriceAdGroupAssetsPerAdGroup = 156,
    AdImageCampaignAssetsPerCampaign = 175,
    AdImageAdGroupAssetsPerAdGroup = 176,
    PageFeedAssetSetsPerCustomer = 157,
    DynamicEducationFeedAssetSetsPerCustomer = 158,
    AssetsPerPageFeedAssetSet = 159,
    AssetsPerDynamicEducationFeedAssetSet = 160,
    DynamicRealEstateAssetSetsPerCustomer = 161,
    AssetsPerDynamicRealEstateAssetSet = 162,
    DynamicCustomAssetSetsPerCustomer = 163,
    AssetsPerDynamicCustomAssetSet = 164,
    DynamicHotelsAndRentalsAssetSetsPerCustomer = 165,
    AssetsPerDynamicHotelsAndRentalsAssetSet = 166,
    DynamicLocalAssetSetsPerCustomer = 167,
    AssetsPerDynamicLocalAssetSet = 168,
    DynamicFlightsAssetSetsPerCustomer = 169,
    AssetsPerDynamicFlightsAssetSet = 170,
    DynamicTravelAssetSetsPerCustomer = 171,
    AssetsPerDynamicTravelAssetSet = 172,
    DynamicJobsAssetSetsPerCustomer = 173,
    AssetsPerDynamicJobsAssetSet = 174,
    BusinessNameCampaignAssetsPerCampaign = 179,
    BusinessLogoCampaignAssetsPerCampaign = 180,
    VersionsPerAd = 82,
    UserFeedsPerCustomer = 90,
    SystemFeedsPerCustomer = 91,
    FeedAttributesPerFeed = 92,
    FeedItemsPerCustomer = 94,
    CampaignFeedsPerCustomer = 95,
    BaseCampaignFeedsPerCustomer = 96,
    ExperimentCampaignFeedsPerCustomer = 109,
    AdGroupFeedsPerCustomer = 97,
    BaseAdGroupFeedsPerCustomer = 98,
    ExperimentAdGroupFeedsPerCustomer = 110,
    AdGroupFeedsPerCampaign = 99,
    FeedItemSetsPerCustomer = 100,
    FeedItemsPerFeedItemSet = 101,
    CampaignExperimentsPerCustomer = 112,
    ExperimentArmsPerVideoExperiment = 113,
    OwnedLabelsPerCustomer = 115,
    LabelsPerCampaign = 117,
    LabelsPerAdGroup = 118,
    LabelsPerAdGroupAd = 119,
    LabelsPerAdGroupCriterion = 120,
    TargetCustomersPerLabel = 121,
    KeywordPlansPerUserPerCustomer = 122,
    KeywordPlanAdGroupKeywordsPerKeywordPlan = 123,
    KeywordPlanAdGroupsPerKeywordPlan = 124,
    KeywordPlanNegativeKeywordsPerKeywordPlan = 125,
    KeywordPlanCampaignsPerKeywordPlan = 126,
    ConversionActionsPerCustomer = 128,
    BatchJobOperationsPerJob = 130,
    BatchJobsPerCustomer = 131,
    HotelCheckInDateRangeBidModifiersPerAdGroup = 132,
    SharedSetsPerAccountForAccountLevelNegativeKeywords = 177,
    AccountLevelNegativeKeywordsPerSharedSet = 178,
    EnabledAssetPerHotelPropertyAssetSet = 181,
    EnabledHotelPropertyAssetLinksPerAssetGroup = 182,
    BrandsPerSharedSet = 183,
    EnabledBrandListCriteriaPerCampaign = 184,
    SharedSetsPerAccountForBrand = 185,
    LookalikeUserListsPerCustomer = 186,

}

Resource limit type.

Variants

Unspecified = 0

No value has been specified.

Unknown = 1

Used for return value only. Represents an unclassified operation unknown in this version.

CampaignsPerCustomer = 2

Number of ENABLED and PAUSED campaigns per customer.

BaseCampaignsPerCustomer = 3

Number of ENABLED and PAUSED base campaigns per customer.

ExperimentCampaignsPerCustomer = 105

Number of ENABLED and PAUSED experiment campaigns per customer.

HotelCampaignsPerCustomer = 4

Number of ENABLED and PAUSED Hotel campaigns per customer.

SmartShoppingCampaignsPerCustomer = 5

Number of ENABLED and PAUSED Smart Shopping campaigns per customer.

AdGroupsPerCampaign = 6

Number of ENABLED ad groups per campaign.

AdGroupsPerShoppingCampaign = 8

Number of ENABLED ad groups per Shopping campaign.

AdGroupsPerHotelCampaign = 9

Number of ENABLED ad groups per Hotel campaign.

ReportingAdGroupsPerLocalCampaign = 10

Number of ENABLED reporting ad groups per local campaign.

ReportingAdGroupsPerAppCampaign = 11

Number of ENABLED reporting ad groups per App campaign. It includes app campaign and app campaign for engagement.

ManagedAdGroupsPerSmartCampaign = 52

Number of ENABLED managed ad groups per smart campaign.

AdGroupCriteriaPerCustomer = 12

Number of ENABLED ad group criteria per customer. An ad group criterion is considered as ENABLED if:

it’s not REMOVED

its ad group is not REMOVED

its campaign is not REMOVED.

BaseAdGroupCriteriaPerCustomer = 13

Number of ad group criteria across all base campaigns for a customer.

ExperimentAdGroupCriteriaPerCustomer = 107

Number of ad group criteria across all experiment campaigns for a customer.

AdGroupCriteriaPerCampaign = 14

Number of ENABLED ad group criteria per campaign. An ad group criterion is considered as ENABLED if:

it’s not REMOVED

its ad group is not REMOVED.

CampaignCriteriaPerCustomer = 15

Number of ENABLED campaign criteria per customer.

BaseCampaignCriteriaPerCustomer = 16

Number of ENABLED campaign criteria across all base campaigns for a customer.

ExperimentCampaignCriteriaPerCustomer = 108

Number of ENABLED campaign criteria across all experiment campaigns for a customer.

WebpageCriteriaPerCustomer = 17

Number of ENABLED webpage criteria per customer, including campaign level and ad group level.

BaseWebpageCriteriaPerCustomer = 18

Number of ENABLED webpage criteria across all base campaigns for a customer.

ExperimentWebpageCriteriaPerCustomer = 19

Meximum number of ENABLED webpage criteria across all experiment campaigns for a customer.

CombinedAudienceCriteriaPerAdGroup = 20

Number of combined audience criteria per ad group.

CustomerNegativePlacementCriteriaPerCustomer = 21

Limit for placement criterion type group in customer negative criterion.

CustomerNegativeYoutubeChannelCriteriaPerCustomer = 22

Limit for YouTube TV channels in customer negative criterion.

CriteriaPerAdGroup = 23

Number of ENABLED criteria per ad group.

ListingGroupsPerAdGroup = 24

Number of listing group criteria per ad group.

ExplicitlySharedBudgetsPerCustomer = 25

Number of ENABLED explicitly shared budgets per customer.

ImplicitlySharedBudgetsPerCustomer = 26

Number of ENABLED implicitly shared budgets per customer.

CombinedAudienceCriteriaPerCampaign = 27

Number of combined audience criteria per campaign.

NegativeKeywordsPerCampaign = 28

Number of negative keywords per campaign.

NegativePlacementsPerCampaign = 29

Number of excluded campaign criteria in placement dimension, for example, placement, mobile application, YouTube channel, etc. The API criterion type is NOT limited to placement only, and this does not include exclusions at the ad group or other levels.

GeoTargetsPerCampaign = 30

Number of geo targets per campaign.

NegativeIpBlocksPerCampaign = 32

Number of negative IP blocks per campaign.

ProximitiesPerCampaign = 33

Number of proximity targets per campaign.

ListingScopesPerShoppingCampaign = 34

Number of listing scopes per Shopping campaign.

ListingScopesPerNonShoppingCampaign = 35

Number of listing scopes per non-Shopping campaign.

NegativeKeywordsPerSharedSet = 36

Number of criteria per negative keyword shared set.

NegativePlacementsPerSharedSet = 37

Number of criteria per negative placement shared set.

SharedSetsPerCustomerForTypeDefault = 40

Default number of shared sets allowed per type per customer.

SharedSetsPerCustomerForNegativePlacementListLower = 41

Number of shared sets of negative placement list type for a manager customer.

HotelAdvanceBookingWindowBidModifiersPerAdGroup = 44

Number of hotel_advance_booking_window bid modifiers per ad group.

BiddingStrategiesPerCustomer = 45

Number of ENABLED shared bidding strategies per customer.

BasicUserListsPerCustomer = 47

Number of open basic user lists per customer.

LogicalUserListsPerCustomer = 48

Number of open logical user lists per customer.

RuleBasedUserListsPerCustomer = 153

Number of open rule based user lists per customer.

BaseAdGroupAdsPerCustomer = 53

Number of ENABLED and PAUSED ad group ads across all base campaigns for a customer.

ExperimentAdGroupAdsPerCustomer = 54

Number of ENABLED and PAUSED ad group ads across all experiment campaigns for a customer.

AdGroupAdsPerCampaign = 55

Number of ENABLED and PAUSED ad group ads per campaign.

TextAndOtherAdsPerAdGroup = 56

Number of ENABLED ads per ad group that do not fall in to other buckets. Includes text and many other types.

ImageAdsPerAdGroup = 57

Number of ENABLED image ads per ad group.

ShoppingSmartAdsPerAdGroup = 58

Number of ENABLED shopping smart ads per ad group.

ResponsiveSearchAdsPerAdGroup = 59

Number of ENABLED responsive search ads per ad group.

AppAdsPerAdGroup = 60

Number of ENABLED app ads per ad group.

AppEngagementAdsPerAdGroup = 61

Number of ENABLED app engagement ads per ad group.

LocalAdsPerAdGroup = 62

Number of ENABLED local ads per ad group.

VideoAdsPerAdGroup = 63

Number of ENABLED video ads per ad group.

LeadFormCampaignAssetsPerCampaign = 143

Number of ENABLED lead form CampaignAssets per campaign.

PromotionCustomerAssetsPerCustomer = 79

Number of ENABLED promotion CustomerAssets per customer.

PromotionCampaignAssetsPerCampaign = 80

Number of ENABLED promotion CampaignAssets per campaign.

PromotionAdGroupAssetsPerAdGroup = 81

Number of ENABLED promotion AdGroupAssets per ad group.

CalloutCustomerAssetsPerCustomer = 134

Number of ENABLED callout CustomerAssets per customer.

CalloutCampaignAssetsPerCampaign = 135

Number of ENABLED callout CampaignAssets per campaign.

CalloutAdGroupAssetsPerAdGroup = 136

Number of ENABLED callout AdGroupAssets per ad group.

SitelinkCustomerAssetsPerCustomer = 137

Number of ENABLED sitelink CustomerAssets per customer.

SitelinkCampaignAssetsPerCampaign = 138

Number of ENABLED sitelink CampaignAssets per campaign.

SitelinkAdGroupAssetsPerAdGroup = 139

Number of ENABLED sitelink AdGroupAssets per ad group.

StructuredSnippetCustomerAssetsPerCustomer = 140

Number of ENABLED structured snippet CustomerAssets per customer.

StructuredSnippetCampaignAssetsPerCampaign = 141

Number of ENABLED structured snippet CampaignAssets per campaign.

StructuredSnippetAdGroupAssetsPerAdGroup = 142

Number of ENABLED structured snippet AdGroupAssets per ad group.

MobileAppCustomerAssetsPerCustomer = 144

Number of ENABLED mobile app CustomerAssets per customer.

MobileAppCampaignAssetsPerCampaign = 145

Number of ENABLED mobile app CampaignAssets per campaign.

MobileAppAdGroupAssetsPerAdGroup = 146

Number of ENABLED mobile app AdGroupAssets per ad group.

HotelCalloutCustomerAssetsPerCustomer = 147

Number of ENABLED hotel callout CustomerAssets per customer.

HotelCalloutCampaignAssetsPerCampaign = 148

Number of ENABLED hotel callout CampaignAssets per campaign.

HotelCalloutAdGroupAssetsPerAdGroup = 149

Number of ENABLED hotel callout AdGroupAssets per ad group.

CallCustomerAssetsPerCustomer = 150

Number of ENABLED call CustomerAssets per customer.

CallCampaignAssetsPerCampaign = 151

Number of ENABLED call CampaignAssets per campaign.

CallAdGroupAssetsPerAdGroup = 152

Number of ENABLED call AdGroupAssets per ad group.

PriceCustomerAssetsPerCustomer = 154

Number of ENABLED price CustomerAssets per customer.

PriceCampaignAssetsPerCampaign = 155

Number of ENABLED price CampaignAssets per campaign.

PriceAdGroupAssetsPerAdGroup = 156

Number of ENABLED price AdGroupAssets per ad group.

AdImageCampaignAssetsPerCampaign = 175

Number of ENABLED ad image CampaignAssets per campaign.

AdImageAdGroupAssetsPerAdGroup = 176

Number of ENABLED ad image AdGroupAssets per ad group.

PageFeedAssetSetsPerCustomer = 157

Number of ENABLED page feed asset sets per customer.

DynamicEducationFeedAssetSetsPerCustomer = 158

Number of ENABLED dynamic education feed asset sets per customer.

AssetsPerPageFeedAssetSet = 159

Number of ENABLED assets per page feed asset set.

AssetsPerDynamicEducationFeedAssetSet = 160

Number of ENABLED assets per dynamic education asset set.

DynamicRealEstateAssetSetsPerCustomer = 161

Number of ENABLED dynamic real estate asset sets per customer.

AssetsPerDynamicRealEstateAssetSet = 162

Number of ENABLED assets per dynamic real estate asset set.

DynamicCustomAssetSetsPerCustomer = 163

Number of ENABLED dynamic custom asset sets per customer.

AssetsPerDynamicCustomAssetSet = 164

Number of ENABLED assets per dynamic custom asset set.

DynamicHotelsAndRentalsAssetSetsPerCustomer = 165

Number of ENABLED dynamic hotels and rentals asset sets per customer.

AssetsPerDynamicHotelsAndRentalsAssetSet = 166

Number of ENABLED assets per dynamic hotels and rentals asset set.

DynamicLocalAssetSetsPerCustomer = 167

Number of ENABLED dynamic local asset sets per customer.

AssetsPerDynamicLocalAssetSet = 168

Number of ENABLED assets per dynamic local asset set.

DynamicFlightsAssetSetsPerCustomer = 169

Number of ENABLED dynamic flights asset sets per customer.

AssetsPerDynamicFlightsAssetSet = 170

Number of ENABLED assets per dynamic flights asset set.

DynamicTravelAssetSetsPerCustomer = 171

Number of ENABLED dynamic travel asset sets per customer.

AssetsPerDynamicTravelAssetSet = 172

Number of ENABLED assets per dynamic travel asset set.

DynamicJobsAssetSetsPerCustomer = 173

Number of ENABLED dynamic jobs asset sets per customer.

AssetsPerDynamicJobsAssetSet = 174

Number of ENABLED assets per dynamic jobs asset set.

BusinessNameCampaignAssetsPerCampaign = 179

Number of ENABLED business name CampaignAssets per campaign.

BusinessLogoCampaignAssetsPerCampaign = 180

Number of ENABLED business logo CampaignAssets per campaign.

VersionsPerAd = 82

Number of versions per ad.

UserFeedsPerCustomer = 90

Number of ENABLED user feeds per customer.

SystemFeedsPerCustomer = 91

Number of ENABLED system feeds per customer.

FeedAttributesPerFeed = 92

Number of feed attributes per feed.

FeedItemsPerCustomer = 94

Number of ENABLED feed items per customer.

CampaignFeedsPerCustomer = 95

Number of ENABLED campaign feeds per customer.

BaseCampaignFeedsPerCustomer = 96

Number of ENABLED campaign feeds across all base campaigns for a customer.

ExperimentCampaignFeedsPerCustomer = 109

Number of ENABLED campaign feeds across all experiment campaigns for a customer.

AdGroupFeedsPerCustomer = 97

Number of ENABLED ad group feeds per customer.

BaseAdGroupFeedsPerCustomer = 98

Number of ENABLED ad group feeds across all base campaigns for a customer.

ExperimentAdGroupFeedsPerCustomer = 110

Number of ENABLED ad group feeds across all experiment campaigns for a customer.

AdGroupFeedsPerCampaign = 99

Number of ENABLED ad group feeds per campaign.

FeedItemSetsPerCustomer = 100

Number of ENABLED feed items per customer.

FeedItemsPerFeedItemSet = 101

Number of feed items per feed item set.

CampaignExperimentsPerCustomer = 112

Number of ENABLED campaign experiments per customer.

ExperimentArmsPerVideoExperiment = 113

Number of video experiment arms per experiment.

OwnedLabelsPerCustomer = 115

Number of owned labels per customer.

LabelsPerCampaign = 117

Number of applied labels per campaign.

LabelsPerAdGroup = 118

Number of applied labels per ad group.

LabelsPerAdGroupAd = 119

Number of applied labels per ad group ad.

LabelsPerAdGroupCriterion = 120

Number of applied labels per ad group criterion.

TargetCustomersPerLabel = 121

Number of customers with a single label applied.

KeywordPlansPerUserPerCustomer = 122

Number of ENABLED keyword plans per user per customer. The limit is applied per <user, customer> pair because by default a plan is private to a user of a customer. Each user of a customer has their own independent limit.

KeywordPlanAdGroupKeywordsPerKeywordPlan = 123

Number of keyword plan ad group keywords per keyword plan.

KeywordPlanAdGroupsPerKeywordPlan = 124

Number of keyword plan ad groups per keyword plan.

KeywordPlanNegativeKeywordsPerKeywordPlan = 125

Number of keyword plan negative keywords (both campaign and ad group) per keyword plan.

KeywordPlanCampaignsPerKeywordPlan = 126

Number of keyword plan campaigns per keyword plan.

ConversionActionsPerCustomer = 128

Number of ENABLED conversion actions per customer.

BatchJobOperationsPerJob = 130

Number of operations in a single batch job.

BatchJobsPerCustomer = 131

Number of PENDING or ENABLED batch jobs per customer.

HotelCheckInDateRangeBidModifiersPerAdGroup = 132

Number of hotel check-in date range bid modifiers per ad agroup.

SharedSetsPerAccountForAccountLevelNegativeKeywords = 177

Number of shared sets of type ACCOUNT_LEVEL_NEGATIVE_KEYWORDS per account.

AccountLevelNegativeKeywordsPerSharedSet = 178

Number of keywords per ACCOUNT_LEVEL_NEGATIVE_KEYWORDS shared set.

EnabledAssetPerHotelPropertyAssetSet = 181

Maximum number of asset per hotel property asset set.

EnabledHotelPropertyAssetLinksPerAssetGroup = 182

Maximum number of enabled hotel property assets per asset group.

BrandsPerSharedSet = 183

Number of criteria per brand shared set.

EnabledBrandListCriteriaPerCampaign = 184

Number of active brand list criteria per campaign.

SharedSetsPerAccountForBrand = 185

Maximum number of shared sets of brand type for an account.

LookalikeUserListsPerCustomer = 186

Maximum number of lookalike lists per customer.

Implementations

impl ResourceLimitType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ResourceLimitType.

pub fn from_i32(value: i32) -> Option<ResourceLimitType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ResourceLimitType, or None if value is not a valid variant.

impl ResourceLimitType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.