Enum google_api_proto::google::ads::googleads::v16::errors::experiment_error_enum::ExperimentError

#[repr(i32)]
pub enum ExperimentError {

    Unspecified = 0,
    Unknown = 1,
    CannotSetStartDateInPast = 2,
    EndDateBeforeStartDate = 3,
    StartDateTooFarInFuture = 4,
    DuplicateExperimentName = 5,
    CannotModifyRemovedExperiment = 6,
    StartDateAlreadyPassed = 7,
    CannotSetEndDateInPast = 8,
    CannotSetStatusToRemoved = 9,
    CannotModifyPastEndDate = 10,
    InvalidStatus = 11,
    InvalidCampaignChannelType = 12,
    OverlappingMembersAndDateRange = 13,
    InvalidTrialArmTrafficSplit = 14,
    TrafficSplitOverlapping = 15,
    SumTrialArmTrafficUnequalsToTrialTrafficSplitDenominator = 16,
    CannotModifyTrafficSplitAfterStart = 17,
    ExperimentNotFound = 18,
    ExperimentNotYetStarted = 19,
    CannotHaveMultipleControlArms = 20,
    InDesignCampaignsNotSet = 21,
    CannotSetStatusToGraduated = 22,
    CannotCreateExperimentCampaignWithSharedBudget = 23,
    CannotCreateExperimentCampaignWithCustomBudget = 24,
    StatusTransitionInvalid = 25,
    DuplicateExperimentCampaignName = 26,
    CannotRemoveInCreationExperiment = 27,
    CannotAddCampaignWithDeprecatedAdTypes = 28,
    CannotEnableSyncForUnsupportedExperimentType = 29,
    InvalidDurationForAnExperiment = 30,

}

Enum describing possible experiment errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

CannotSetStartDateInPast = 2

The start date of an experiment cannot be set in the past. Use a start date in the future.

EndDateBeforeStartDate = 3

The end date of an experiment is before its start date. Use an end date after the start date.

StartDateTooFarInFuture = 4

The start date of an experiment is too far in the future. Use a start date no more than 1 year in the future.

DuplicateExperimentName = 5

The experiment has the same name as an existing active experiment.

CannotModifyRemovedExperiment = 6

Experiments can only be modified when they are ENABLED.

StartDateAlreadyPassed = 7

The start date of an experiment cannot be modified if the existing start date has already passed.

CannotSetEndDateInPast = 8

The end date of an experiment cannot be set in the past.

CannotSetStatusToRemoved = 9

The status of an experiment cannot be set to REMOVED.

CannotModifyPastEndDate = 10

The end date of an expired experiment cannot be modified.

InvalidStatus = 11

The status is invalid.

InvalidCampaignChannelType = 12

Experiment arm contains campaigns with invalid advertising channel type.

OverlappingMembersAndDateRange = 13

A pair of trials share members and have overlapping date ranges.

InvalidTrialArmTrafficSplit = 14

Experiment arm contains invalid traffic split.

TrafficSplitOverlapping = 15

Experiment contains trial arms with overlapping traffic split.

SumTrialArmTrafficUnequalsToTrialTrafficSplitDenominator = 16

The total traffic split of trial arms is not equal to 100.

CannotModifyTrafficSplitAfterStart = 17

Traffic split related settings (like traffic share bounds) can’t be modified after the experiment has started.

ExperimentNotFound = 18

The experiment could not be found.

ExperimentNotYetStarted = 19

Experiment has not begun.

CannotHaveMultipleControlArms = 20

The experiment cannot have more than one control arm.

InDesignCampaignsNotSet = 21

The experiment doesn’t set in-design campaigns.

CannotSetStatusToGraduated = 22

Clients must use the graduate action to graduate experiments and cannot set the status to GRADUATED directly.

CannotCreateExperimentCampaignWithSharedBudget = 23

Cannot use shared budget on base campaign when scheduling an experiment.

CannotCreateExperimentCampaignWithCustomBudget = 24

Cannot use custom budget on base campaign when scheduling an experiment.

StatusTransitionInvalid = 25

Invalid status transition.

DuplicateExperimentCampaignName = 26

The experiment campaign name conflicts with a pre-existing campaign.

CannotRemoveInCreationExperiment = 27

Cannot remove in creation experiments.

CannotAddCampaignWithDeprecatedAdTypes = 28

Cannot add campaign with deprecated ad types. Deprecated ad types: ENHANCED_DISPLAY, GALLERY, GMAIL, KEYWORDLESS, TEXT.

CannotEnableSyncForUnsupportedExperimentType = 29

Sync can only be enabled for supported experiment types. Supported experiment types: SEARCH_CUSTOM, DISPLAY_CUSTOM, DISPLAY_AUTOMATED_BIDDING_STRATEGY, SEARCH_AUTOMATED_BIDDING_STRATEGY.

InvalidDurationForAnExperiment = 30

Experiment length cannot be longer than max length.

Implementations

impl ExperimentError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ExperimentError.

pub fn from_i32(value: i32) -> Option<ExperimentError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ExperimentError, or None if value is not a valid variant.

impl ExperimentError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.