Enum google_api_proto::google::ads::googleads::v16::errors::extension_setting_error_enum::ExtensionSettingError

#[repr(i32)]
pub enum ExtensionSettingError {

    Unspecified = 0,
    Unknown = 1,
    ExtensionsRequired = 2,
    FeedTypeExtensionTypeMismatch = 3,
    InvalidFeedType = 4,
    InvalidFeedTypeForCustomerExtensionSetting = 5,
    CannotChangeFeedItemOnCreate = 6,
    CannotUpdateNewlyCreatedExtension = 7,
    NoExistingAdGroupExtensionSettingForType = 8,
    NoExistingCampaignExtensionSettingForType = 9,
    NoExistingCustomerExtensionSettingForType = 10,
    AdGroupExtensionSettingAlreadyExists = 11,
    CampaignExtensionSettingAlreadyExists = 12,
    CustomerExtensionSettingAlreadyExists = 13,
    AdGroupFeedAlreadyExistsForPlaceholderType = 14,
    CampaignFeedAlreadyExistsForPlaceholderType = 15,
    CustomerFeedAlreadyExistsForPlaceholderType = 16,
    ValueOutOfRange = 17,
    CannotSetFieldWithFinalUrls = 18,
    FinalUrlsNotSet = 19,
    InvalidPhoneNumber = 20,
    PhoneNumberNotSupportedForCountry = 21,
    CarrierSpecificShortNumberNotAllowed = 22,
    PremiumRateNumberNotAllowed = 23,
    DisallowedNumberType = 24,
    InvalidDomesticPhoneNumberFormat = 25,
    VanityPhoneNumberNotAllowed = 26,
    InvalidCountryCode = 27,
    InvalidCallConversionTypeId = 28,
    CustomerNotInAllowlistForCalltracking = 69,
    CalltrackingNotSupportedForCountry = 30,
    InvalidAppId = 31,
    QuotesInReviewExtensionSnippet = 32,
    HyphensInReviewExtensionSnippet = 33,
    ReviewExtensionSourceNotEligible = 34,
    SourceNameInReviewExtensionText = 35,
    MissingField = 36,
    InconsistentCurrencyCodes = 37,
    PriceExtensionHasDuplicatedHeaders = 38,
    PriceItemHasDuplicatedHeaderAndDescription = 39,
    PriceExtensionHasTooFewItems = 40,
    PriceExtensionHasTooManyItems = 41,
    UnsupportedValue = 42,
    InvalidDevicePreference = 43,
    InvalidScheduleEnd = 45,
    DateTimeMustBeInAccountTimeZone = 47,
    OverlappingSchedulesNotAllowed = 48,
    ScheduleEndNotAfterStart = 49,
    TooManySchedulesPerDay = 50,
    DuplicateExtensionFeedItemEdit = 51,
    InvalidSnippetsHeader = 52,
    PhoneNumberNotSupportedWithCalltrackingForCountry = 53,
    CampaignTargetingMismatch = 54,
    CannotOperateOnRemovedFeed = 55,
    ExtensionTypeRequired = 56,
    IncompatibleUnderlyingMatchingFunction = 57,
    StartDateAfterEndDate = 58,
    InvalidPriceFormat = 59,
    PromotionInvalidTime = 60,
    PromotionCannotSetPercentDiscountAndMoneyDiscount = 61,
    PromotionCannotSetPromotionCodeAndOrdersOverAmount = 62,
    TooManyDecimalPlacesSpecified = 63,
    InvalidLanguageCode = 64,
    UnsupportedLanguage = 65,
    CustomerConsentForCallRecordingRequired = 66,
    ExtensionSettingUpdateIsANoop = 67,
    DisallowedText = 68,

}

Enum describing possible extension setting errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

ExtensionsRequired = 2

A platform restriction was provided without input extensions or existing extensions.

FeedTypeExtensionTypeMismatch = 3

The provided feed type does not correspond to the provided extensions.

InvalidFeedType = 4

The provided feed type cannot be used.

InvalidFeedTypeForCustomerExtensionSetting = 5

The provided feed type cannot be used at the customer level.

CannotChangeFeedItemOnCreate = 6

Cannot change a feed item field on a CREATE operation.

CannotUpdateNewlyCreatedExtension = 7

Cannot update an extension that is not already in this setting.

NoExistingAdGroupExtensionSettingForType = 8

There is no existing AdGroupExtensionSetting for this type.

NoExistingCampaignExtensionSettingForType = 9

There is no existing CampaignExtensionSetting for this type.

NoExistingCustomerExtensionSettingForType = 10

There is no existing CustomerExtensionSetting for this type.

AdGroupExtensionSettingAlreadyExists = 11

The AdGroupExtensionSetting already exists. UPDATE should be used to modify the existing AdGroupExtensionSetting.

CampaignExtensionSettingAlreadyExists = 12

The CampaignExtensionSetting already exists. UPDATE should be used to modify the existing CampaignExtensionSetting.

CustomerExtensionSettingAlreadyExists = 13

The CustomerExtensionSetting already exists. UPDATE should be used to modify the existing CustomerExtensionSetting.

AdGroupFeedAlreadyExistsForPlaceholderType = 14

An active ad group feed already exists for this place holder type.

CampaignFeedAlreadyExistsForPlaceholderType = 15

An active campaign feed already exists for this place holder type.

CustomerFeedAlreadyExistsForPlaceholderType = 16

An active customer feed already exists for this place holder type.

ValueOutOfRange = 17

Value is not within the accepted range.

CannotSetFieldWithFinalUrls = 18

Cannot simultaneously set specified field with final urls.

FinalUrlsNotSet = 19

Must set field with final urls.

InvalidPhoneNumber = 20

Phone number for a call extension is invalid.

PhoneNumberNotSupportedForCountry = 21

Phone number for a call extension is not supported for the given country code.

CarrierSpecificShortNumberNotAllowed = 22

A carrier specific number in short format is not allowed for call extensions.

PremiumRateNumberNotAllowed = 23

Premium rate numbers are not allowed for call extensions.

DisallowedNumberType = 24

Phone number type for a call extension is not allowed.

InvalidDomesticPhoneNumberFormat = 25

Phone number for a call extension does not meet domestic format requirements.

VanityPhoneNumberNotAllowed = 26

Vanity phone numbers (for example, those including letters) are not allowed for call extensions.

InvalidCountryCode = 27

Country code provided for a call extension is invalid.

InvalidCallConversionTypeId = 28

Call conversion type id provided for a call extension is invalid.

CustomerNotInAllowlistForCalltracking = 69

For a call extension, the customer is not on the allow-list for call tracking.

CalltrackingNotSupportedForCountry = 30

Call tracking is not supported for the given country for a call extension.

InvalidAppId = 31

App id provided for an app extension is invalid.

QuotesInReviewExtensionSnippet = 32

Quotation marks present in the review text for a review extension.

HyphensInReviewExtensionSnippet = 33

Hyphen character present in the review text for a review extension.

ReviewExtensionSourceNotEligible = 34

A blocked review source name or url was provided for a review extension.

SourceNameInReviewExtensionText = 35

Review source name should not be found in the review text.

MissingField = 36

Field must be set.

InconsistentCurrencyCodes = 37

Inconsistent currency codes.

PriceExtensionHasDuplicatedHeaders = 38

Price extension cannot have duplicated headers.

PriceItemHasDuplicatedHeaderAndDescription = 39

Price item cannot have duplicated header and description.

PriceExtensionHasTooFewItems = 40

Price extension has too few items

PriceExtensionHasTooManyItems = 41

Price extension has too many items

UnsupportedValue = 42

The input value is not currently supported.

InvalidDevicePreference = 43

Unknown or unsupported device preference.

InvalidScheduleEnd = 45

Invalid feed item schedule end time (for example, endHour = 24 and endMinute != 0).

DateTimeMustBeInAccountTimeZone = 47

Date time zone does not match the account’s time zone.

OverlappingSchedulesNotAllowed = 48

Overlapping feed item schedule times (for example, 7-10AM and 8-11AM) are not allowed.

ScheduleEndNotAfterStart = 49

Feed item schedule end time must be after start time.

TooManySchedulesPerDay = 50

There are too many feed item schedules per day.

DuplicateExtensionFeedItemEdit = 51

Cannot edit the same extension feed item more than once in the same request.

InvalidSnippetsHeader = 52

Invalid structured snippet header.

PhoneNumberNotSupportedWithCalltrackingForCountry = 53

Phone number with call tracking enabled is not supported for the specified country.

CampaignTargetingMismatch = 54

The targeted adgroup must belong to the targeted campaign.

CannotOperateOnRemovedFeed = 55

The feed used by the ExtensionSetting is removed and cannot be operated on. Remove the ExtensionSetting to allow a new one to be created using an active feed.

ExtensionTypeRequired = 56

The ExtensionFeedItem type is required for this operation.

IncompatibleUnderlyingMatchingFunction = 57

The matching function that links the extension feed to the customer, campaign, or ad group is not compatible with the ExtensionSetting services.

StartDateAfterEndDate = 58

Start date must be before end date.

InvalidPriceFormat = 59

Input price is not in a valid format.

PromotionInvalidTime = 60

The promotion time is invalid.

PromotionCannotSetPercentDiscountAndMoneyDiscount = 61

Cannot set both percent discount and money discount fields.

PromotionCannotSetPromotionCodeAndOrdersOverAmount = 62

Cannot set both promotion code and orders over amount fields.

TooManyDecimalPlacesSpecified = 63

This field has too many decimal places specified.

InvalidLanguageCode = 64

The language code is not valid.

UnsupportedLanguage = 65

The language is not supported.

CustomerConsentForCallRecordingRequired = 66

Customer hasn’t consented for call recording, which is required for adding/updating call extensions. See https://support.google.com/google-ads/answer/7412639.

ExtensionSettingUpdateIsANoop = 67

The UPDATE operation does not specify any fields other than the resource name in the update mask.

DisallowedText = 68

The extension contains text which has been prohibited on policy grounds.

Implementations

impl ExtensionSettingError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ExtensionSettingError.

pub fn from_i32(value: i32) -> Option<ExtensionSettingError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ExtensionSettingError, or None if value is not a valid variant.

impl ExtensionSettingError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.