Struct google_api_proto::google::ads::googleads::v16::resources::AccountBudget

pub struct AccountBudget {

    pub resource_name: String,
    pub id: Option<i64>,
    pub billing_setup: Option<String>,
    pub status: i32,
    pub name: Option<String>,
    pub proposed_start_date_time: Option<String>,
    pub approved_start_date_time: Option<String>,
    pub total_adjustments_micros: i64,
    pub amount_served_micros: i64,
    pub purchase_order_number: Option<String>,
    pub notes: Option<String>,
    pub pending_proposal: Option<PendingAccountBudgetProposal>,
    pub proposed_end_time: Option<ProposedEndTime>,
    pub approved_end_time: Option<ApprovedEndTime>,
    pub proposed_spending_limit: Option<ProposedSpendingLimit>,
    pub approved_spending_limit: Option<ApprovedSpendingLimit>,
    pub adjusted_spending_limit: Option<AdjustedSpendingLimit>,

}

An account-level budget. It contains information about the budget itself, as well as the most recently approved changes to the budget and proposed changes that are pending approval. The proposed changes that are pending approval, if any, are found in ‘pending_proposal’. Effective details about the budget are found in fields prefixed ‘approved_’, ‘adjusted_’ and those without a prefix. Since some effective details may differ from what the user had originally requested (for example, spending limit), these differences are juxtaposed through ‘proposed_’, ‘approved_’, and possibly ‘adjusted_’ fields.

This resource is mutated using AccountBudgetProposal and cannot be mutated directly. A budget may have at most one pending proposal at any given time. It is read through pending_proposal.

Once approved, a budget may be subject to adjustments, such as credit adjustments. Adjustments create differences between the ‘approved’ and ‘adjusted’ fields, which would otherwise be identical.

Fields
resource_name: String

Output only. The resource name of the account-level budget. AccountBudget resource names have the form:

customers/{customer_id}/accountBudgets/{account_budget_id}

id: Option<i64>

Output only. The ID of the account-level budget.

billing_setup: Option<String>

Output only. The resource name of the billing setup associated with this account-level budget. BillingSetup resource names have the form:

customers/{customer_id}/billingSetups/{billing_setup_id}

status: i32

Output only. The status of this account-level budget.

name: Option<String>

Output only. The name of the account-level budget.

proposed_start_date_time: Option<String>

Output only. The proposed start time of the account-level budget in yyyy-MM-dd HH:mm:ss format. If a start time type of NOW was proposed, this is the time of request.

approved_start_date_time: Option<String>

Output only. The approved start time of the account-level budget in yyyy-MM-dd HH:mm:ss format.

For example, if a new budget is approved after the proposed start time, the approved start time is the time of approval.

total_adjustments_micros: i64

Output only. The total adjustments amount.

An example of an adjustment is courtesy credits.

amount_served_micros: i64

Output only. The value of Ads that have been served, in micros.

This includes overdelivery costs, in which case a credit might be automatically applied to the budget (see total_adjustments_micros).

purchase_order_number: Option<String>

Output only. A purchase order number is a value that helps users reference this budget in their monthly invoices.

notes: Option<String>

Output only. Notes associated with the budget.

pending_proposal: Option<PendingAccountBudgetProposal>

Output only. The pending proposal to modify this budget, if applicable.

proposed_end_time: Option<ProposedEndTime>

The proposed end time of the account-level budget.

approved_end_time: Option<ApprovedEndTime>

The approved end time of the account-level budget.

For example, if a budget’s end time is updated and the proposal is approved after the proposed end time, the approved end time is the time of approval.

proposed_spending_limit: Option<ProposedSpendingLimit>

The proposed spending limit.

approved_spending_limit: Option<ApprovedSpendingLimit>

The approved spending limit.

For example, if the amount already spent by the account exceeds the proposed spending limit at the time the proposal is approved, the approved spending limit is set to the amount already spent.

adjusted_spending_limit: Option<AdjustedSpendingLimit>

The spending limit after adjustments have been applied. Adjustments are stored in total_adjustments_micros.

This value has the final say on how much the account is allowed to spend.

Implementations

impl AccountBudget

pub fn status(&self) -> AccountBudgetStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: AccountBudgetStatus)

Sets status to the provided enum value.

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn billing_setup(&self) -> &str

Returns the value of billing_setup, or the default value if billing_setup is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn proposed_start_date_time(&self) -> &str

Returns the value of proposed_start_date_time, or the default value if proposed_start_date_time is unset.

pub fn approved_start_date_time(&self) -> &str

Returns the value of approved_start_date_time, or the default value if approved_start_date_time is unset.

pub fn purchase_order_number(&self) -> &str

Returns the value of purchase_order_number, or the default value if purchase_order_number is unset.

pub fn notes(&self) -> &str

Returns the value of notes, or the default value if notes is unset.