Struct google_api_proto::google::ads::googleads::v16::resources::Invoice

pub struct Invoice {

    pub resource_name: String,
    pub id: Option<String>,
    pub type: i32,
    pub billing_setup: Option<String>,
    pub payments_account_id: Option<String>,
    pub payments_profile_id: Option<String>,
    pub issue_date: Option<String>,
    pub due_date: Option<String>,
    pub service_date_range: Option<DateRange>,
    pub currency_code: Option<String>,
    pub adjustments_subtotal_amount_micros: i64,
    pub adjustments_tax_amount_micros: i64,
    pub adjustments_total_amount_micros: i64,
    pub regulatory_costs_subtotal_amount_micros: i64,
    pub regulatory_costs_tax_amount_micros: i64,
    pub regulatory_costs_total_amount_micros: i64,
    pub export_charge_subtotal_amount_micros: Option<i64>,
    pub export_charge_tax_amount_micros: Option<i64>,
    pub export_charge_total_amount_micros: Option<i64>,
    pub subtotal_amount_micros: Option<i64>,
    pub tax_amount_micros: Option<i64>,
    pub total_amount_micros: Option<i64>,
    pub corrected_invoice: Option<String>,
    pub replaced_invoices: Vec<String>,
    pub pdf_url: Option<String>,
    pub account_budget_summaries: Vec<AccountBudgetSummary>,
    pub account_summaries: Vec<AccountSummary>,

}

An invoice. All invoice information is snapshotted to match the PDF invoice. For invoices older than the launch of InvoiceService, the snapshotted information may not match the PDF invoice.

Fields
resource_name: String

Output only. The resource name of the invoice. Multiple customers can share a given invoice, so multiple resource names may point to the same invoice. Invoice resource names have the form:

customers/{customer_id}/invoices/{invoice_id}

id: Option<String>

Output only. The ID of the invoice. It appears on the invoice PDF as “Invoice number”.

type: i32

Output only. The type of invoice.

billing_setup: Option<String>

Output only. The resource name of this invoice’s billing setup.

customers/{customer_id}/billingSetups/{billing_setup_id}

payments_account_id: Option<String>

Output only. A 16 digit ID used to identify the payments account associated with the billing setup, for example, “1234-5678-9012-3456”. It appears on the invoice PDF as “Billing Account Number”.

payments_profile_id: Option<String>

Output only. A 12 digit ID used to identify the payments profile associated with the billing setup, for example, “1234-5678-9012”. It appears on the invoice PDF as “Billing ID”.

issue_date: Option<String>

Output only. The issue date in yyyy-mm-dd format. It appears on the invoice PDF as either “Issue date” or “Invoice date”.

due_date: Option<String>

Output only. The due date in yyyy-mm-dd format.

service_date_range: Option<DateRange>

Output only. The service period date range of this invoice. The end date is inclusive.

currency_code: Option<String>

Output only. The currency code. All costs are returned in this currency. A subset of the currency codes derived from the ISO 4217 standard is supported.

adjustments_subtotal_amount_micros: i64

Output only. The pretax subtotal amount of invoice level adjustments, in micros.

adjustments_tax_amount_micros: i64

Output only. The sum of taxes on the invoice level adjustments, in micros.

adjustments_total_amount_micros: i64

Output only. The total amount of invoice level adjustments, in micros.

regulatory_costs_subtotal_amount_micros: i64

Output only. The pretax subtotal amount of invoice level regulatory costs, in micros.

regulatory_costs_tax_amount_micros: i64

Output only. The sum of taxes on the invoice level regulatory costs, in micros.

regulatory_costs_total_amount_micros: i64

Output only. The total amount of invoice level regulatory costs, in micros.

export_charge_subtotal_amount_micros: Option<i64>

Output only. The pretax subtotal amount of invoice level export charges, in micros.

export_charge_tax_amount_micros: Option<i64>

Output only. The sum of taxes on the invoice level export charges, in micros.

export_charge_total_amount_micros: Option<i64>

Output only. The total amount of invoice level export charges, in micros.

subtotal_amount_micros: Option<i64>

Output only. The pretax subtotal amount, in micros. This is equal to the sum of the AccountBudgetSummary subtotal amounts and Invoice.adjustments_subtotal_amount_micros.

tax_amount_micros: Option<i64>

Output only. The sum of all taxes on the invoice, in micros. This equals the sum of the AccountBudgetSummary tax amounts, plus taxes not associated with a specific account budget.

total_amount_micros: Option<i64>

Output only. The total amount, in micros. This equals the sum of Invoice.subtotal_amount_micros, Invoice.tax_amount_micros, Invoice.regulatory_costs_subtotal_amount_micros, and Invoice.export_charge_subtotal_amount_micros (which is separated into a separate line item starting with V14.1).

corrected_invoice: Option<String>

Output only. The resource name of the original invoice corrected, wrote off, or canceled by this invoice, if applicable. If corrected_invoice is set, replaced_invoices will not be set. Invoice resource names have the form:

customers/{customer_id}/invoices/{invoice_id}

replaced_invoices: Vec<String>

Output only. The resource name of the original invoice(s) being rebilled or replaced by this invoice, if applicable. There might be multiple replaced invoices due to invoice consolidation. The replaced invoices may not belong to the same payments account. If replaced_invoices is set, corrected_invoice will not be set. Invoice resource names have the form:

customers/{customer_id}/invoices/{invoice_id}

pdf_url: Option<String>

Output only. The URL to a PDF copy of the invoice. Users need to pass in their OAuth token to request the PDF with this URL.

account_budget_summaries: Vec<AccountBudgetSummary>

Output only. The list of summarized account budget information associated with this invoice.

account_summaries: Vec<AccountSummary>

Output only. The list of summarized account information associated with this invoice.

Implementations

impl Invoice

pub fn type(&self) -> InvoiceType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: InvoiceType)

Sets type to the provided enum value.

pub fn id(&self) -> &str

Returns the value of id, or the default value if id is unset.

pub fn billing_setup(&self) -> &str

Returns the value of billing_setup, or the default value if billing_setup is unset.

pub fn payments_account_id(&self) -> &str

Returns the value of payments_account_id, or the default value if payments_account_id is unset.

pub fn payments_profile_id(&self) -> &str

Returns the value of payments_profile_id, or the default value if payments_profile_id is unset.

pub fn issue_date(&self) -> &str

Returns the value of issue_date, or the default value if issue_date is unset.

pub fn due_date(&self) -> &str

Returns the value of due_date, or the default value if due_date is unset.

pub fn currency_code(&self) -> &str

Returns the value of currency_code, or the default value if currency_code is unset.

pub fn subtotal_amount_micros(&self) -> i64

Returns the value of subtotal_amount_micros, or the default value if subtotal_amount_micros is unset.

pub fn tax_amount_micros(&self) -> i64

Returns the value of tax_amount_micros, or the default value if tax_amount_micros is unset.

pub fn total_amount_micros(&self) -> i64

Returns the value of total_amount_micros, or the default value if total_amount_micros is unset.

pub fn corrected_invoice(&self) -> &str

Returns the value of corrected_invoice, or the default value if corrected_invoice is unset.

pub fn pdf_url(&self) -> &str

Returns the value of pdf_url, or the default value if pdf_url is unset.

pub fn export_charge_subtotal_amount_micros(&self) -> i64

Returns the value of export_charge_subtotal_amount_micros, or the default value if export_charge_subtotal_amount_micros is unset.

pub fn export_charge_tax_amount_micros(&self) -> i64

Returns the value of export_charge_tax_amount_micros, or the default value if export_charge_tax_amount_micros is unset.

pub fn export_charge_total_amount_micros(&self) -> i64

Returns the value of export_charge_total_amount_micros, or the default value if export_charge_total_amount_micros is unset.