Enum google_api_proto::google::ads::googleads::v16::services::mutate_operation::Operation

pub enum Operation {

    AdGroupAdLabelOperation(AdGroupAdLabelOperation),
    AdGroupAdOperation(AdGroupAdOperation),
    AdGroupAssetOperation(AdGroupAssetOperation),
    AdGroupBidModifierOperation(AdGroupBidModifierOperation),
    AdGroupCriterionCustomizerOperation(AdGroupCriterionCustomizerOperation),
    AdGroupCriterionLabelOperation(AdGroupCriterionLabelOperation),
    AdGroupCriterionOperation(AdGroupCriterionOperation),
    AdGroupCustomizerOperation(AdGroupCustomizerOperation),
    AdGroupExtensionSettingOperation(AdGroupExtensionSettingOperation),
    AdGroupFeedOperation(AdGroupFeedOperation),
    AdGroupLabelOperation(AdGroupLabelOperation),
    AdGroupOperation(AdGroupOperation),
    AdOperation(AdOperation),
    AdParameterOperation(AdParameterOperation),
    AssetOperation(AssetOperation),
    AssetGroupAssetOperation(AssetGroupAssetOperation),
    AssetGroupListingGroupFilterOperation(AssetGroupListingGroupFilterOperation),
    AssetGroupSignalOperation(AssetGroupSignalOperation),
    AssetGroupOperation(AssetGroupOperation),
    AssetSetAssetOperation(AssetSetAssetOperation),
    AssetSetOperation(AssetSetOperation),
    AudienceOperation(AudienceOperation),
    BiddingDataExclusionOperation(BiddingDataExclusionOperation),
    BiddingSeasonalityAdjustmentOperation(BiddingSeasonalityAdjustmentOperation),
    BiddingStrategyOperation(BiddingStrategyOperation),
    CampaignAssetOperation(CampaignAssetOperation),
    CampaignAssetSetOperation(CampaignAssetSetOperation),
    CampaignBidModifierOperation(CampaignBidModifierOperation),
    CampaignBudgetOperation(CampaignBudgetOperation),
    CampaignConversionGoalOperation(CampaignConversionGoalOperation),
    CampaignCriterionOperation(CampaignCriterionOperation),
    CampaignCustomizerOperation(CampaignCustomizerOperation),
    CampaignDraftOperation(CampaignDraftOperation),
    CampaignExtensionSettingOperation(CampaignExtensionSettingOperation),
    CampaignFeedOperation(CampaignFeedOperation),
    CampaignGroupOperation(CampaignGroupOperation),
    CampaignLabelOperation(CampaignLabelOperation),
    CampaignOperation(CampaignOperation),
    CampaignSharedSetOperation(CampaignSharedSetOperation),
    ConversionActionOperation(ConversionActionOperation),
    ConversionCustomVariableOperation(ConversionCustomVariableOperation),
    ConversionGoalCampaignConfigOperation(ConversionGoalCampaignConfigOperation),
    ConversionValueRuleOperation(ConversionValueRuleOperation),
    ConversionValueRuleSetOperation(ConversionValueRuleSetOperation),
    CustomConversionGoalOperation(CustomConversionGoalOperation),
    CustomerAssetOperation(CustomerAssetOperation),
    CustomerConversionGoalOperation(CustomerConversionGoalOperation),
    CustomerCustomizerOperation(CustomerCustomizerOperation),
    CustomerExtensionSettingOperation(CustomerExtensionSettingOperation),
    CustomerFeedOperation(CustomerFeedOperation),
    CustomerLabelOperation(CustomerLabelOperation),
    CustomerNegativeCriterionOperation(CustomerNegativeCriterionOperation),
    CustomerOperation(CustomerOperation),
    CustomizerAttributeOperation(CustomizerAttributeOperation),
    ExperimentOperation(ExperimentOperation),
    ExperimentArmOperation(ExperimentArmOperation),
    ExtensionFeedItemOperation(ExtensionFeedItemOperation),
    FeedItemOperation(FeedItemOperation),
    FeedItemSetOperation(FeedItemSetOperation),
    FeedItemSetLinkOperation(FeedItemSetLinkOperation),
    FeedItemTargetOperation(FeedItemTargetOperation),
    FeedMappingOperation(FeedMappingOperation),
    FeedOperation(FeedOperation),
    KeywordPlanAdGroupOperation(KeywordPlanAdGroupOperation),
    KeywordPlanAdGroupKeywordOperation(KeywordPlanAdGroupKeywordOperation),
    KeywordPlanCampaignKeywordOperation(KeywordPlanCampaignKeywordOperation),
    KeywordPlanCampaignOperation(KeywordPlanCampaignOperation),
    KeywordPlanOperation(KeywordPlanOperation),
    LabelOperation(LabelOperation),
    RecommendationSubscriptionOperation(RecommendationSubscriptionOperation),
    RemarketingActionOperation(RemarketingActionOperation),
    SharedCriterionOperation(SharedCriterionOperation),
    SharedSetOperation(SharedSetOperation),
    SmartCampaignSettingOperation(SmartCampaignSettingOperation),
    UserListOperation(UserListOperation),

}

The mutate operation.

Variants

AdGroupAdLabelOperation(AdGroupAdLabelOperation)

An ad group ad label mutate operation.

AdGroupAdOperation(AdGroupAdOperation)

An ad group ad mutate operation.

AdGroupAssetOperation(AdGroupAssetOperation)

An ad group asset mutate operation.

AdGroupBidModifierOperation(AdGroupBidModifierOperation)

An ad group bid modifier mutate operation.

AdGroupCriterionCustomizerOperation(AdGroupCriterionCustomizerOperation)

An ad group criterion customizer mutate operation.

AdGroupCriterionLabelOperation(AdGroupCriterionLabelOperation)

An ad group criterion label mutate operation.

AdGroupCriterionOperation(AdGroupCriterionOperation)

An ad group criterion mutate operation.

AdGroupCustomizerOperation(AdGroupCustomizerOperation)

An ad group customizer mutate operation.

AdGroupExtensionSettingOperation(AdGroupExtensionSettingOperation)

An ad group extension setting mutate operation.

AdGroupFeedOperation(AdGroupFeedOperation)

An ad group feed mutate operation.

AdGroupLabelOperation(AdGroupLabelOperation)

An ad group label mutate operation.

AdGroupOperation(AdGroupOperation)

An ad group mutate operation.

AdOperation(AdOperation)

An ad mutate operation.

AdParameterOperation(AdParameterOperation)

An ad parameter mutate operation.

AssetOperation(AssetOperation)

An asset mutate operation.

AssetGroupAssetOperation(AssetGroupAssetOperation)

An asset group asset mutate operation.

AssetGroupListingGroupFilterOperation(AssetGroupListingGroupFilterOperation)

An asset group listing group filter mutate operation.

AssetGroupSignalOperation(AssetGroupSignalOperation)

An asset group signal mutate operation.

AssetGroupOperation(AssetGroupOperation)

An asset group mutate operation.

AssetSetAssetOperation(AssetSetAssetOperation)

An asset set asset mutate operation.

AssetSetOperation(AssetSetOperation)

An asset set mutate operation.

AudienceOperation(AudienceOperation)

An audience mutate operation.

BiddingDataExclusionOperation(BiddingDataExclusionOperation)

A bidding data exclusion mutate operation.

BiddingSeasonalityAdjustmentOperation(BiddingSeasonalityAdjustmentOperation)

A bidding seasonality adjustment mutate operation.

BiddingStrategyOperation(BiddingStrategyOperation)

A bidding strategy mutate operation.

CampaignAssetOperation(CampaignAssetOperation)

A campaign asset mutate operation.

CampaignAssetSetOperation(CampaignAssetSetOperation)

A campaign asset mutate operation.

CampaignBidModifierOperation(CampaignBidModifierOperation)

A campaign bid modifier mutate operation.

CampaignBudgetOperation(CampaignBudgetOperation)

A campaign budget mutate operation.

CampaignConversionGoalOperation(CampaignConversionGoalOperation)

A campaign conversion goal mutate operation.

CampaignCriterionOperation(CampaignCriterionOperation)

A campaign criterion mutate operation.

CampaignCustomizerOperation(CampaignCustomizerOperation)

A campaign customizer mutate operation.

CampaignDraftOperation(CampaignDraftOperation)

A campaign draft mutate operation.

CampaignExtensionSettingOperation(CampaignExtensionSettingOperation)

A campaign extension setting mutate operation.

CampaignFeedOperation(CampaignFeedOperation)

A campaign feed mutate operation.

CampaignGroupOperation(CampaignGroupOperation)

A campaign group mutate operation.

CampaignLabelOperation(CampaignLabelOperation)

A campaign label mutate operation.

CampaignOperation(CampaignOperation)

A campaign mutate operation.

CampaignSharedSetOperation(CampaignSharedSetOperation)

A campaign shared set mutate operation.

ConversionActionOperation(ConversionActionOperation)

A conversion action mutate operation.

ConversionCustomVariableOperation(ConversionCustomVariableOperation)

A conversion custom variable mutate operation.

ConversionGoalCampaignConfigOperation(ConversionGoalCampaignConfigOperation)

A conversion goal campaign config mutate operation.

ConversionValueRuleOperation(ConversionValueRuleOperation)

A conversion value rule mutate operation.

ConversionValueRuleSetOperation(ConversionValueRuleSetOperation)

A conversion value rule set mutate operation.

CustomConversionGoalOperation(CustomConversionGoalOperation)

A custom conversion goal mutate operation.

CustomerAssetOperation(CustomerAssetOperation)

A customer asset mutate operation.

CustomerConversionGoalOperation(CustomerConversionGoalOperation)

A customer conversion goal mutate operation.

CustomerCustomizerOperation(CustomerCustomizerOperation)

A customer customizer mutate operation.

CustomerExtensionSettingOperation(CustomerExtensionSettingOperation)

A customer extension setting mutate operation.

CustomerFeedOperation(CustomerFeedOperation)

A customer feed mutate operation.

CustomerLabelOperation(CustomerLabelOperation)

A customer label mutate operation.

CustomerNegativeCriterionOperation(CustomerNegativeCriterionOperation)

A customer negative criterion mutate operation.

CustomerOperation(CustomerOperation)

A customer mutate operation.

CustomizerAttributeOperation(CustomizerAttributeOperation)

A customizer attribute mutate operation.

ExperimentOperation(ExperimentOperation)

An experiment mutate operation.

ExperimentArmOperation(ExperimentArmOperation)

An experiment arm mutate operation.

ExtensionFeedItemOperation(ExtensionFeedItemOperation)

An extension feed item mutate operation.

FeedItemOperation(FeedItemOperation)

A feed item mutate operation.

FeedItemSetOperation(FeedItemSetOperation)

A feed item set mutate operation.

FeedItemSetLinkOperation(FeedItemSetLinkOperation)

A feed item set link mutate operation.

FeedItemTargetOperation(FeedItemTargetOperation)

A feed item target mutate operation.

FeedMappingOperation(FeedMappingOperation)

A feed mapping mutate operation.

FeedOperation(FeedOperation)

A feed mutate operation.

KeywordPlanAdGroupOperation(KeywordPlanAdGroupOperation)

A keyword plan ad group operation.

KeywordPlanAdGroupKeywordOperation(KeywordPlanAdGroupKeywordOperation)

A keyword plan ad group keyword operation.

KeywordPlanCampaignKeywordOperation(KeywordPlanCampaignKeywordOperation)

A keyword plan campaign keyword operation.

KeywordPlanCampaignOperation(KeywordPlanCampaignOperation)

A keyword plan campaign operation.

KeywordPlanOperation(KeywordPlanOperation)

A keyword plan operation.

LabelOperation(LabelOperation)

A label mutate operation.

RecommendationSubscriptionOperation(RecommendationSubscriptionOperation)

A recommendation subscription mutate operation.

RemarketingActionOperation(RemarketingActionOperation)

A remarketing action mutate operation.

SharedCriterionOperation(SharedCriterionOperation)

A shared criterion mutate operation.

SharedSetOperation(SharedSetOperation)

A shared set mutate operation.

SmartCampaignSettingOperation(SmartCampaignSettingOperation)

A Smart campaign setting mutate operation.

UserListOperation(UserListOperation)

A user list mutate operation.

Implementations

impl Operation

pub fn encode<B>(&self, buf: &mut B)
where B: BufMut,

Encodes the message to a buffer.

pub fn merge<B>( field: &mut Option<Operation>, tag: u32, wire_type: WireType, buf: &mut B, ctx: DecodeContext ) -> Result<(), DecodeError>
where B: Buf,

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.