Enum google_api_proto::google::ads::googleads::v17::common::listing_dimension_info::Dimension

pub enum Dimension {

    HotelId(HotelIdInfo),
    HotelClass(HotelClassInfo),
    HotelCountryRegion(HotelCountryRegionInfo),
    HotelState(HotelStateInfo),
    HotelCity(HotelCityInfo),
    ProductCategory(ProductCategoryInfo),
    ProductBrand(ProductBrandInfo),
    ProductChannel(ProductChannelInfo),
    ProductChannelExclusivity(ProductChannelExclusivityInfo),
    ProductCondition(ProductConditionInfo),
    ProductCustomAttribute(ProductCustomAttributeInfo),
    ProductItemId(ProductItemIdInfo),
    ProductType(ProductTypeInfo),
    ProductGrouping(ProductGroupingInfo),
    ProductLabels(ProductLabelsInfo),
    ProductLegacyCondition(ProductLegacyConditionInfo),
    ProductTypeFull(ProductTypeFullInfo),
    ActivityId(ActivityIdInfo),
    ActivityRating(ActivityRatingInfo),
    ActivityCountry(ActivityCountryInfo),
    ActivityState(ActivityStateInfo),
    ActivityCity(ActivityCityInfo),
    UnknownListingDimension(UnknownListingDimensionInfo),

}

Dimension of one of the types below is always present.

Variants

HotelId(HotelIdInfo)

Advertiser-specific hotel ID.

HotelClass(HotelClassInfo)

Class of the hotel as a number of stars 1 to 5.

HotelCountryRegion(HotelCountryRegionInfo)

Country or Region the hotel is located in.

HotelState(HotelStateInfo)

State the hotel is located in.

HotelCity(HotelCityInfo)

City the hotel is located in.

ProductCategory(ProductCategoryInfo)

Category of a product offer.

ProductBrand(ProductBrandInfo)

Brand of a product offer.

ProductChannel(ProductChannelInfo)

Locality of a product offer.

ProductChannelExclusivity(ProductChannelExclusivityInfo)

Availability of a product offer.

ProductCondition(ProductConditionInfo)

Condition of a product offer.

ProductCustomAttribute(ProductCustomAttributeInfo)

Custom attribute of a product offer.

ProductItemId(ProductItemIdInfo)

Item id of a product offer.

ProductType(ProductTypeInfo)

Type of a product offer.

ProductGrouping(ProductGroupingInfo)

Grouping of a product offer. This listing dimension is deprecated and it is supported only in Display campaigns.

ProductLabels(ProductLabelsInfo)

Labels of a product offer. This listing dimension is deprecated and it is supported only in Display campaigns.

ProductLegacyCondition(ProductLegacyConditionInfo)

Legacy condition of a product offer. This listing dimension is deprecated and it is supported only in Display campaigns.

ProductTypeFull(ProductTypeFullInfo)

Full type of a product offer. This listing dimension is deprecated and it is supported only in Display campaigns.

ActivityId(ActivityIdInfo)

Advertiser-specific activity ID.

ActivityRating(ActivityRatingInfo)

Rating of the activity as a number 1 to 5, where 5 is the best.

ActivityCountry(ActivityCountryInfo)

The country where the travel activity is available.

ActivityState(ActivityStateInfo)

The state where the travel activity is available.

ActivityCity(ActivityCityInfo)

The city where the travel activity is available.

UnknownListingDimension(UnknownListingDimensionInfo)

Unknown dimension. Set when no other listing dimension is set.

Implementations

impl Dimension

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Dimension>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.