Enum google_api_proto::google::ads::googleads::v17::enums::click_type_enum::ClickType

#[repr(i32)]
pub enum ClickType {

    Unspecified = 0,
    Unknown = 1,
    AppDeeplink = 2,
    Breadcrumbs = 3,
    BroadbandPlan = 4,
    CallTracking = 5,
    Calls = 6,
    ClickOnEngagementAd = 7,
    GetDirections = 8,
    LocationExpansion = 9,
    LocationFormatCall = 10,
    LocationFormatDirections = 11,
    LocationFormatImage = 12,
    LocationFormatLandingPage = 13,
    LocationFormatMap = 14,
    LocationFormatStoreInfo = 15,
    LocationFormatText = 16,
    MobileCallTracking = 17,
    OfferPrints = 18,
    Other = 19,
    ProductExtensionClicks = 20,
    ProductListingAdClicks = 21,
    Sitelinks = 22,
    StoreLocator = 23,
    UrlClicks = 25,
    VideoAppStoreClicks = 26,
    VideoCallToActionClicks = 27,
    VideoCardActionHeadlineClicks = 28,
    VideoEndCapClicks = 29,
    VideoWebsiteClicks = 30,
    VisualSitelinks = 31,
    WirelessPlan = 32,
    ProductListingAdLocal = 33,
    ProductListingAdMultichannelLocal = 34,
    ProductListingAdMultichannelOnline = 35,
    ProductListingAdsCoupon = 36,
    ProductListingAdTransactable = 37,
    ProductAdAppDeeplink = 38,
    ShowcaseAdCategoryLink = 39,
    ShowcaseAdLocalStorefrontLink = 40,
    ShowcaseAdOnlineProductLink = 42,
    ShowcaseAdLocalProductLink = 43,
    PromotionExtension = 44,
    SwipeableGalleryAdHeadline = 45,
    SwipeableGalleryAdSwipes = 46,
    SwipeableGalleryAdSeeMore = 47,
    SwipeableGalleryAdSitelinkOne = 48,
    SwipeableGalleryAdSitelinkTwo = 49,
    SwipeableGalleryAdSitelinkThree = 50,
    SwipeableGalleryAdSitelinkFour = 51,
    SwipeableGalleryAdSitelinkFive = 52,
    HotelPrice = 53,
    PriceExtension = 54,
    HotelBookOnGoogleRoomSelection = 55,
    ShoppingComparisonListing = 56,
    CrossNetwork = 57,
    AdImage = 58,

}

Enumerates Google Ads click types.

Variants

Unspecified = 0

Not specified.

Unknown = 1

The value is unknown in this version.

AppDeeplink = 2

App engagement ad deep link.

Breadcrumbs = 3

Breadcrumbs.

BroadbandPlan = 4

Broadband Plan.

CallTracking = 5

Manually dialed phone calls.

Calls = 6

Phone calls.

ClickOnEngagementAd = 7

Click on engagement ad.

GetDirections = 8

Driving direction.

LocationExpansion = 9

Get location details.

LocationFormatCall = 10

Call.

LocationFormatDirections = 11

Directions.

LocationFormatImage = 12

Image(s).

LocationFormatLandingPage = 13

Go to landing page.

LocationFormatMap = 14

Map.

LocationFormatStoreInfo = 15

Go to store info.

LocationFormatText = 16

Text.

MobileCallTracking = 17

Mobile phone calls.

OfferPrints = 18

Print offer.

Other = 19

Other.

ProductExtensionClicks = 20

Product plusbox offer.

ProductListingAdClicks = 21

Shopping - Product - Online.

Sitelinks = 22

Sitelink.

StoreLocator = 23

Show nearby locations.

UrlClicks = 25

Headline.

VideoAppStoreClicks = 26

App store.

VideoCallToActionClicks = 27

Call-to-Action overlay.

VideoCardActionHeadlineClicks = 28

Cards.

VideoEndCapClicks = 29

End cap.

VideoWebsiteClicks = 30

Website.

VisualSitelinks = 31

Visual Sitelinks.

WirelessPlan = 32

Wireless Plan.

ProductListingAdLocal = 33

Shopping - Product - Local.

ProductListingAdMultichannelLocal = 34

Shopping - Product - MultiChannel Local.

ProductListingAdMultichannelOnline = 35

Shopping - Product - MultiChannel Online.

ProductListingAdsCoupon = 36

Shopping - Product - Coupon.

ProductListingAdTransactable = 37

Shopping - Product - Sell on Google.

ProductAdAppDeeplink = 38

Shopping - Product - App engagement ad deep link.

ShowcaseAdCategoryLink = 39

Shopping - Showcase - Category.

ShowcaseAdLocalStorefrontLink = 40

Shopping - Showcase - Local storefront.

ShowcaseAdOnlineProductLink = 42

Shopping - Showcase - Online product.

ShowcaseAdLocalProductLink = 43

Shopping - Showcase - Local product.

PromotionExtension = 44

Promotion Extension.

SwipeableGalleryAdHeadline = 45

Ad Headline.

SwipeableGalleryAdSwipes = 46

Swipes.

SwipeableGalleryAdSeeMore = 47

See More.

SwipeableGalleryAdSitelinkOne = 48

Sitelink 1.

SwipeableGalleryAdSitelinkTwo = 49

Sitelink 2.

SwipeableGalleryAdSitelinkThree = 50

Sitelink 3.

SwipeableGalleryAdSitelinkFour = 51

Sitelink 4.

SwipeableGalleryAdSitelinkFive = 52

Sitelink 5.

HotelPrice = 53

Hotel price.

PriceExtension = 54

Price Extension.

HotelBookOnGoogleRoomSelection = 55

Book on Google hotel room selection.

ShoppingComparisonListing = 56

Shopping - Comparison Listing.

CrossNetwork = 57

Cross-network. From Performance Max and Discovery Campaigns.

AdImage = 58

AD Images.

Implementations

impl ClickType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ClickType.

pub fn from_i32(value: i32) -> Option<ClickType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ClickType, or None if value is not a valid variant.

impl ClickType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.