Enum google_api_proto::google::ads::googleads::v17::errors::criterion_error_enum::CriterionError

#[repr(i32)]
pub enum CriterionError {

    Unspecified = 0,
    Unknown = 1,
    ConcreteTypeRequired = 2,
    InvalidExcludedCategory = 3,
    InvalidKeywordText = 4,
    KeywordTextTooLong = 5,
    KeywordHasTooManyWords = 6,
    KeywordHasInvalidChars = 7,
    InvalidPlacementUrl = 8,
    InvalidUserList = 9,
    InvalidUserInterest = 10,
    InvalidFormatForPlacementUrl = 11,
    PlacementUrlIsTooLong = 12,
    PlacementUrlHasIllegalChar = 13,
    PlacementUrlHasMultipleSitesInLine = 14,
    PlacementIsNotAvailableForTargetingOrExclusion = 15,
    InvalidTopicPath = 16,
    InvalidYoutubeChannelId = 17,
    InvalidYoutubeVideoId = 18,
    YoutubeVerticalChannelDeprecated = 19,
    YoutubeDemographicChannelDeprecated = 20,
    YoutubeUrlUnsupported = 21,
    CannotExcludeCriteriaType = 22,
    CannotAddCriteriaType = 23,
    CannotExcludeSimilarUserList = 26,
    CannotAddClosedUserList = 27,
    CannotAddDisplayOnlyListsToSearchOnlyCampaigns = 28,
    CannotAddDisplayOnlyListsToSearchCampaigns = 29,
    CannotAddDisplayOnlyListsToShoppingCampaigns = 30,
    CannotAddUserInterestsToSearchCampaigns = 31,
    CannotSetBidsOnCriterionTypeInSearchCampaigns = 32,
    CannotAddUrlsToCriterionTypeForCampaignType = 33,
    InvalidCombinedAudience = 122,
    InvalidCustomAffinity = 96,
    InvalidCustomIntent = 97,
    InvalidCustomAudience = 121,
    InvalidIpAddress = 34,
    InvalidIpFormat = 35,
    InvalidMobileApp = 36,
    InvalidMobileAppCategory = 37,
    InvalidCriterionId = 38,
    CannotTargetCriterion = 39,
    CannotTargetObsoleteCriterion = 40,
    CriterionIdAndTypeMismatch = 41,
    InvalidProximityRadius = 42,
    InvalidProximityRadiusUnits = 43,
    InvalidStreetaddressLength = 44,
    InvalidCitynameLength = 45,
    InvalidRegioncodeLength = 46,
    InvalidRegionnameLength = 47,
    InvalidPostalcodeLength = 48,
    InvalidCountryCode = 49,
    InvalidLatitude = 50,
    InvalidLongitude = 51,
    ProximityGeopointAndAddressBothCannotBeNull = 52,
    InvalidProximityAddress = 53,
    InvalidUserDomainName = 54,
    CriterionParameterTooLong = 55,
    AdScheduleTimeIntervalsOverlap = 56,
    AdScheduleIntervalCannotSpanMultipleDays = 57,
    AdScheduleInvalidTimeInterval = 58,
    AdScheduleExceededIntervalsPerDayLimit = 59,
    AdScheduleCriterionIdMismatchingFields = 60,
    CannotBidModifyCriterionType = 61,
    CannotBidModifyCriterionCampaignOptedOut = 62,
    CannotBidModifyNegativeCriterion = 63,
    BidModifierAlreadyExists = 64,
    FeedIdNotAllowed = 65,
    AccountIneligibleForCriteriaType = 66,
    CriteriaTypeInvalidForBiddingStrategy = 67,
    CannotExcludeCriterion = 68,
    CannotRemoveCriterion = 69,
    InvalidProductBiddingCategory = 76,
    MissingShoppingSetting = 77,
    InvalidMatchingFunction = 78,
    LocationFilterNotAllowed = 79,
    InvalidFeedForLocationFilter = 98,
    LocationFilterInvalid = 80,
    CannotSetGeoTargetConstantsWithFeedItemSets = 123,
    CannotSetBothAssetSetAndFeed = 140,
    CannotSetFeedOrFeedItemSetsForCustomer = 142,
    CannotSetAssetSetFieldForCustomer = 150,
    CannotSetGeoTargetConstantsWithAssetSets = 143,
    CannotSetAssetSetsWithFeedItemSets = 144,
    InvalidLocationGroupAssetSet = 141,
    InvalidLocationGroupRadius = 124,
    InvalidLocationGroupRadiusUnit = 125,
    CannotAttachCriteriaAtCampaignAndAdgroup = 81,
    HotelLengthOfStayOverlapsWithExistingCriterion = 82,
    HotelAdvanceBookingWindowOverlapsWithExistingCriterion = 83,
    FieldIncompatibleWithNegativeTargeting = 84,
    InvalidWebpageCondition = 85,
    InvalidWebpageConditionUrl = 86,
    WebpageConditionUrlCannotBeEmpty = 87,
    WebpageConditionUrlUnsupportedProtocol = 88,
    WebpageConditionUrlCannotBeIpAddress = 89,
    WebpageConditionUrlDomainNotConsistentWithCampaignSetting = 90,
    WebpageConditionUrlCannotBePublicSuffix = 91,
    WebpageConditionUrlInvalidPublicSuffix = 92,
    WebpageConditionUrlValueTrackValueNotSupported = 93,
    WebpageCriterionUrlEqualsCanHaveOnlyOneCondition = 94,
    WebpageCriterionNotSupportedOnNonDsaAdGroup = 95,
    CannotTargetUserListForSmartDisplayCampaigns = 99,
    CannotTargetPlacementsForSearchCampaigns = 126,
    ListingScopeTooManyDimensionTypes = 100,
    ListingScopeTooManyInOperators = 101,
    ListingScopeInOperatorNotSupported = 102,
    DuplicateListingDimensionType = 103,
    DuplicateListingDimensionValue = 104,
    CannotSetBidsOnListingGroupSubdivision = 105,
    InvalidListingGroupHierarchy = 106,
    ListingGroupUnitCannotHaveChildren = 107,
    ListingGroupSubdivisionRequiresOthersCase = 108,
    ListingGroupRequiresSameDimensionTypeAsSiblings = 109,
    ListingGroupAlreadyExists = 110,
    ListingGroupDoesNotExist = 111,
    ListingGroupCannotBeRemoved = 112,
    InvalidListingGroupType = 113,
    ListingGroupAddMayOnlyUseTempId = 114,
    ListingScopeTooLong = 115,
    ListingScopeTooManyDimensions = 116,
    ListingGroupTooLong = 117,
    ListingGroupTreeTooDeep = 118,
    InvalidListingDimension = 119,
    InvalidListingDimensionType = 120,
    AdvertiserNotOnAllowlistForCombinedAudienceOnDisplay = 127,
    CannotTargetRemovedCombinedAudience = 128,
    InvalidCombinedAudienceId = 129,
    CannotTargetRemovedCustomAudience = 130,
    HotelCheckInDateRangeOverlapsWithExistingCriterion = 131,
    HotelCheckInDateRangeStartDateTooEarly = 132,
    HotelCheckInDateRangeEndDateTooLate = 133,
    HotelCheckInDateRangeReversed = 134,
    BroadMatchModifierKeywordNotAllowed = 135,
    OneAudienceAllowedPerAssetGroup = 136,
    AudienceNotEligibleForCampaignType = 137,
    AudienceNotAllowedToAttachWhenAudienceGroupedSetToFalse = 138,
    CannotTargetCustomerMatchUserList = 139,
    NegativeKeywordSharedSetDoesNotExist = 145,
    CannotAddRemovedNegativeKeywordSharedSet = 146,
    CannotHaveMultipleNegativeKeywordListPerAccount = 147,
    CustomerCannotAddCriterionOfThisType = 149,
    CannotTargetSimilarUserList = 151,
    CannotAddAudienceSegmentCriterionWhenAudienceGroupedIsSet = 152,
    OneAudienceAllowedPerAdGroup = 153,
    InvalidDetailedDemographic = 154,
    CannotRecognizeBrand = 155,
    BrandSharedSetDoesNotExist = 156,
    CannotAddRemovedBrandSharedSet = 157,
    OnlyExclusionBrandListAllowedForCampaignType = 158,

}

Enum describing possible criterion errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

ConcreteTypeRequired = 2

Concrete type of criterion is required for CREATE and UPDATE operations.

InvalidExcludedCategory = 3

The category requested for exclusion is invalid.

InvalidKeywordText = 4

Invalid keyword criteria text.

KeywordTextTooLong = 5

Keyword text should be less than 80 chars.

KeywordHasTooManyWords = 6

Keyword text has too many words.

KeywordHasInvalidChars = 7

Keyword text has invalid characters or symbols.

InvalidPlacementUrl = 8

Invalid placement URL.

InvalidUserList = 9

Invalid user list criterion.

InvalidUserInterest = 10

Invalid user interest criterion.

InvalidFormatForPlacementUrl = 11

Placement URL has wrong format.

PlacementUrlIsTooLong = 12

Placement URL is too long.

PlacementUrlHasIllegalChar = 13

Indicates the URL contains an illegal character.

PlacementUrlHasMultipleSitesInLine = 14

Indicates the URL contains multiple comma separated URLs.

PlacementIsNotAvailableForTargetingOrExclusion = 15

Indicates the domain is blocked.

InvalidTopicPath = 16

Invalid topic path.

InvalidYoutubeChannelId = 17

The YouTube Channel Id is invalid.

InvalidYoutubeVideoId = 18

The YouTube Video Id is invalid.

YoutubeVerticalChannelDeprecated = 19

Indicates the placement is a YouTube vertical channel, which is no longer supported.

YoutubeDemographicChannelDeprecated = 20

Indicates the placement is a YouTube demographic channel, which is no longer supported.

YoutubeUrlUnsupported = 21

YouTube urls are not supported in Placement criterion. Use YouTubeChannel and YouTubeVideo criterion instead.

CannotExcludeCriteriaType = 22

Criteria type can not be excluded by the customer, like AOL account type cannot target site type criteria.

CannotAddCriteriaType = 23

Criteria type can not be targeted.

CannotExcludeSimilarUserList = 26

Not allowed to exclude similar user list.

CannotAddClosedUserList = 27

Not allowed to target a closed user list.

CannotAddDisplayOnlyListsToSearchOnlyCampaigns = 28

Not allowed to add display only UserLists to search only campaigns.

CannotAddDisplayOnlyListsToSearchCampaigns = 29

Not allowed to add display only UserLists to search plus campaigns.

CannotAddDisplayOnlyListsToShoppingCampaigns = 30

Not allowed to add display only UserLists to shopping campaigns.

CannotAddUserInterestsToSearchCampaigns = 31

Not allowed to add User interests to search only campaigns.

CannotSetBidsOnCriterionTypeInSearchCampaigns = 32

Not allowed to set bids for this criterion type in search campaigns

CannotAddUrlsToCriterionTypeForCampaignType = 33

Final URLs, URL Templates and CustomParameters cannot be set for the criterion types of Gender, AgeRange, UserList, Placement, MobileApp, and MobileAppCategory in search campaigns and shopping campaigns.

InvalidCombinedAudience = 122

Invalid combined audience criterion.

InvalidCustomAffinity = 96

Invalid custom affinity criterion.

InvalidCustomIntent = 97

Invalid custom intent criterion.

InvalidCustomAudience = 121

Invalid custom audience criterion.

InvalidIpAddress = 34

IP address is not valid.

InvalidIpFormat = 35

IP format is not valid.

InvalidMobileApp = 36

Mobile application is not valid.

InvalidMobileAppCategory = 37

Mobile application category is not valid.

InvalidCriterionId = 38

The CriterionId does not exist or is of the incorrect type.

CannotTargetCriterion = 39

The Criterion is not allowed to be targeted.

CannotTargetObsoleteCriterion = 40

The criterion is not allowed to be targeted as it is deprecated.

CriterionIdAndTypeMismatch = 41

The CriterionId is not valid for the type.

InvalidProximityRadius = 42

Distance for the radius for the proximity criterion is invalid.

InvalidProximityRadiusUnits = 43

Units for the distance for the radius for the proximity criterion is invalid.

InvalidStreetaddressLength = 44

Street address in the address is not valid.

InvalidCitynameLength = 45

City name in the address is not valid.

InvalidRegioncodeLength = 46

Region code in the address is not valid.

InvalidRegionnameLength = 47

Region name in the address is not valid.

InvalidPostalcodeLength = 48

Postal code in the address is not valid.

InvalidCountryCode = 49

Country code in the address is not valid.

InvalidLatitude = 50

Latitude for the GeoPoint is not valid.

InvalidLongitude = 51

Longitude for the GeoPoint is not valid.

ProximityGeopointAndAddressBothCannotBeNull = 52

The Proximity input is not valid. Both address and geoPoint cannot be null.

InvalidProximityAddress = 53

The Proximity address cannot be geocoded to a valid lat/long.

InvalidUserDomainName = 54

User domain name is not valid.

CriterionParameterTooLong = 55

Length of serialized criterion parameter exceeded size limit.

AdScheduleTimeIntervalsOverlap = 56

Time interval in the AdSchedule overlaps with another AdSchedule.

AdScheduleIntervalCannotSpanMultipleDays = 57

AdSchedule time interval cannot span multiple days.

AdScheduleInvalidTimeInterval = 58

AdSchedule time interval specified is invalid, endTime cannot be earlier than startTime.

AdScheduleExceededIntervalsPerDayLimit = 59

The number of AdSchedule entries in a day exceeds the limit.

AdScheduleCriterionIdMismatchingFields = 60

CriteriaId does not match the interval of the AdSchedule specified.

CannotBidModifyCriterionType = 61

Cannot set bid modifier for this criterion type.

CannotBidModifyCriterionCampaignOptedOut = 62

Cannot bid modify criterion, since it is opted out of the campaign.

CannotBidModifyNegativeCriterion = 63

Cannot set bid modifier for a negative criterion.

BidModifierAlreadyExists = 64

Bid Modifier already exists. Use SET operation to update.

FeedIdNotAllowed = 65

Feed Id is not allowed in these Location Groups.

AccountIneligibleForCriteriaType = 66

The account may not use the requested criteria type. For example, some accounts are restricted to keywords only.

CriteriaTypeInvalidForBiddingStrategy = 67

The requested criteria type cannot be used with campaign or ad group bidding strategy.

CannotExcludeCriterion = 68

The Criterion is not allowed to be excluded.

CannotRemoveCriterion = 69

The criterion is not allowed to be removed. For example, we cannot remove any of the device criterion.

InvalidProductBiddingCategory = 76

Bidding categories do not form a valid path in the Shopping bidding category taxonomy.

MissingShoppingSetting = 77

ShoppingSetting must be added to the campaign before ProductScope criteria can be added.

InvalidMatchingFunction = 78

Matching function is invalid.

LocationFilterNotAllowed = 79

Filter parameters not allowed for location groups targeting.

InvalidFeedForLocationFilter = 98

Feed not found, or the feed is not an enabled location feed.

LocationFilterInvalid = 80

Given location filter parameter is invalid for location groups targeting.

CannotSetGeoTargetConstantsWithFeedItemSets = 123

Cannot set geo target constants and feed item sets at the same time.

CannotSetBothAssetSetAndFeed = 140

Cannot set both assetset and feed at the same time.

CannotSetFeedOrFeedItemSetsForCustomer = 142

Cannot set feed or feed item sets for Customer.

CannotSetAssetSetFieldForCustomer = 150

Cannot set AssetSet criteria for customer.

CannotSetGeoTargetConstantsWithAssetSets = 143

Cannot set geo target constants and asset sets at the same time.

CannotSetAssetSetsWithFeedItemSets = 144

Cannot set asset sets and feed item sets at the same time.

InvalidLocationGroupAssetSet = 141

The location group asset set id is invalid

InvalidLocationGroupRadius = 124

The location group radius is in the range but not at the valid increment.

InvalidLocationGroupRadiusUnit = 125

The location group radius unit is invalid.

CannotAttachCriteriaAtCampaignAndAdgroup = 81

Criteria type cannot be associated with a campaign and its ad group(s) simultaneously.

HotelLengthOfStayOverlapsWithExistingCriterion = 82

Range represented by hotel length of stay’s min nights and max nights overlaps with an existing criterion.

HotelAdvanceBookingWindowOverlapsWithExistingCriterion = 83

Range represented by hotel advance booking window’s min days and max days overlaps with an existing criterion.

FieldIncompatibleWithNegativeTargeting = 84

The field is not allowed to be set when the negative field is set to true, for example, we don’t allow bids in negative ad group or campaign criteria.

InvalidWebpageCondition = 85

The combination of operand and operator in webpage condition is invalid.

InvalidWebpageConditionUrl = 86

The URL of webpage condition is invalid.

WebpageConditionUrlCannotBeEmpty = 87

The URL of webpage condition cannot be empty or contain white space.

WebpageConditionUrlUnsupportedProtocol = 88

The URL of webpage condition contains an unsupported protocol.

WebpageConditionUrlCannotBeIpAddress = 89

The URL of webpage condition cannot be an IP address.

WebpageConditionUrlDomainNotConsistentWithCampaignSetting = 90

The domain of the URL is not consistent with the domain in campaign setting.

WebpageConditionUrlCannotBePublicSuffix = 91

The URL of webpage condition cannot be a public suffix itself.

WebpageConditionUrlInvalidPublicSuffix = 92

The URL of webpage condition has an invalid public suffix.

WebpageConditionUrlValueTrackValueNotSupported = 93

Value track parameter is not supported in webpage condition URL.

WebpageCriterionUrlEqualsCanHaveOnlyOneCondition = 94

Only one URL-EQUALS webpage condition is allowed in a webpage criterion and it cannot be combined with other conditions.

WebpageCriterionNotSupportedOnNonDsaAdGroup = 95

A webpage criterion cannot be added to a non-DSA ad group.

CannotTargetUserListForSmartDisplayCampaigns = 99

Cannot add positive user list criteria in Smart Display campaigns.

CannotTargetPlacementsForSearchCampaigns = 126

Cannot add positive placement criterion types in search campaigns.

ListingScopeTooManyDimensionTypes = 100

Listing scope contains too many dimension types.

ListingScopeTooManyInOperators = 101

Listing scope has too many IN operators.

ListingScopeInOperatorNotSupported = 102

Listing scope contains IN operator on an unsupported dimension type.

DuplicateListingDimensionType = 103

There are dimensions with duplicate dimension type.

DuplicateListingDimensionValue = 104

There are dimensions with duplicate dimension value.

CannotSetBidsOnListingGroupSubdivision = 105

Listing group SUBDIVISION nodes cannot have bids.

InvalidListingGroupHierarchy = 106

Ad group is invalid due to the listing groups it contains.

ListingGroupUnitCannotHaveChildren = 107

Listing group unit cannot have children.

ListingGroupSubdivisionRequiresOthersCase = 108

Subdivided listing groups must have an “others” case.

ListingGroupRequiresSameDimensionTypeAsSiblings = 109

Dimension type of listing group must be the same as that of its siblings.

ListingGroupAlreadyExists = 110

Listing group cannot be added to the ad group because it already exists.

ListingGroupDoesNotExist = 111

Listing group referenced in the operation was not found in the ad group.

ListingGroupCannotBeRemoved = 112

Recursive removal failed because listing group subdivision is being created or modified in this request.

InvalidListingGroupType = 113

Listing group type is not allowed for specified ad group criterion type.

ListingGroupAddMayOnlyUseTempId = 114

Listing group in an ADD operation specifies a non temporary criterion id.

ListingScopeTooLong = 115

The combined length of dimension values of the Listing scope criterion is too long.

ListingScopeTooManyDimensions = 116

Listing scope contains too many dimensions.

ListingGroupTooLong = 117

The combined length of dimension values of the Listing group criterion is too long.

ListingGroupTreeTooDeep = 118

Listing group tree is too deep.

InvalidListingDimension = 119

Listing dimension is invalid (for example, dimension contains illegal value, dimension type is represented with wrong class, etc). Listing dimension value can not contain “==” or “&+”.

InvalidListingDimensionType = 120

Listing dimension type is either invalid for campaigns of this type or cannot be used in the current context. BIDDING_CATEGORY_Lx and PRODUCT_TYPE_Lx dimensions must be used in ascending order of their levels: L1, L2, L3, L4, L5… The levels must be specified sequentially and start from L1. Furthermore, an “others” Listing group cannot be subdivided with a dimension of the same type but of a higher level (“others” BIDDING_CATEGORY_L3 can be subdivided with BRAND but not with BIDDING_CATEGORY_L4).

AdvertiserNotOnAllowlistForCombinedAudienceOnDisplay = 127

Customer is not on allowlist for composite audience in display campaigns.

CannotTargetRemovedCombinedAudience = 128

Cannot target on a removed combined audience.

InvalidCombinedAudienceId = 129

Combined audience ID is invalid.

CannotTargetRemovedCustomAudience = 130

Can not target removed combined audience.

HotelCheckInDateRangeOverlapsWithExistingCriterion = 131

Range represented by hotel check-in date’s start date and end date overlaps with an existing criterion.

HotelCheckInDateRangeStartDateTooEarly = 132

Start date is earlier than earliest allowed value of yesterday UTC.

HotelCheckInDateRangeEndDateTooLate = 133

End date later is than latest allowed day of 330 days in the future UTC.

HotelCheckInDateRangeReversed = 134

Start date is after end date.

BroadMatchModifierKeywordNotAllowed = 135

Broad match modifier (BMM) keywords can no longer be created. See https://ads-developers.googleblog.com/2021/06/broad-match-modifier-upcoming-changes.html.

OneAudienceAllowedPerAssetGroup = 136

Only one audience is allowed in an asset group.

AudienceNotEligibleForCampaignType = 137

Audience is not supported for the specified campaign type.

AudienceNotAllowedToAttachWhenAudienceGroupedSetToFalse = 138

Audience is not allowed to attach when use_audience_grouped bit is set to false.

CannotTargetCustomerMatchUserList = 139

Targeting is not allowed for Customer Match lists as per Customer Match policy. See https://support.google.com/google-ads/answer/6299717.

NegativeKeywordSharedSetDoesNotExist = 145

Cannot create a negative keyword list criterion with a shared set that does not exist.

CannotAddRemovedNegativeKeywordSharedSet = 146

Cannot create a negative keyword list with deleted shared set.

CannotHaveMultipleNegativeKeywordListPerAccount = 147

Can only have one Negative Keyword List per account.

CustomerCannotAddCriterionOfThisType = 149

Only allowlisted customers can add criteria of this type.

CannotTargetSimilarUserList = 151

Targeting for Similar audiences is not supported, since this feature has been deprecated. See https://support.google.com/google-ads/answer/12463119 to learn more.

CannotAddAudienceSegmentCriterionWhenAudienceGroupedIsSet = 152

Audience segment criteria cannot be added when use_audience_grouped bit is set.

OneAudienceAllowedPerAdGroup = 153

Only one audience is allowed in an ad group.

InvalidDetailedDemographic = 154

Invalid detailed demographics criterion.

CannotRecognizeBrand = 155

The brand criteria has a brand input that is not recognized as a valid brand.

BrandSharedSetDoesNotExist = 156

The brand_list.shared_set_id references a shared set that does not exist.

CannotAddRemovedBrandSharedSet = 157

Cannot create a brand list with deleted shared set.

OnlyExclusionBrandListAllowedForCampaignType = 158

Brand list can only be negatively targeted for the campaign type.

Implementations

impl CriterionError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of CriterionError.

pub fn from_i32(value: i32) -> Option<CriterionError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a CriterionError, or None if value is not a valid variant.

impl CriterionError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.