Enum google_api_proto::google::ads::googleads::v17::errors::error_code::ErrorCode

pub enum ErrorCode {

    RequestError(i32),
    BiddingStrategyError(i32),
    UrlFieldError(i32),
    ListOperationError(i32),
    QueryError(i32),
    MutateError(i32),
    FieldMaskError(i32),
    AuthorizationError(i32),
    InternalError(i32),
    QuotaError(i32),
    AdError(i32),
    AdGroupError(i32),
    CampaignBudgetError(i32),
    CampaignError(i32),
    VideoCampaignError(i32),
    AuthenticationError(i32),
    AdGroupCriterionCustomizerError(i32),
    AdGroupCriterionError(i32),
    AdGroupCustomizerError(i32),
    AdCustomizerError(i32),
    AdGroupAdError(i32),
    AdSharingError(i32),
    AdxError(i32),
    AssetError(i32),
    AssetGroupAssetError(i32),
    AssetGroupListingGroupFilterError(i32),
    AssetGroupError(i32),
    AssetSetAssetError(i32),
    AssetSetLinkError(i32),
    AssetSetError(i32),
    BiddingError(i32),
    CampaignCriterionError(i32),
    CampaignConversionGoalError(i32),
    CampaignCustomizerError(i32),
    CollectionSizeError(i32),
    ConversionGoalCampaignConfigError(i32),
    CountryCodeError(i32),
    CriterionError(i32),
    CustomConversionGoalError(i32),
    CustomerCustomizerError(i32),
    CustomerError(i32),
    CustomizerAttributeError(i32),
    DateError(i32),
    DateRangeError(i32),
    DistinctError(i32),
    FeedAttributeReferenceError(i32),
    FunctionError(i32),
    FunctionParsingError(i32),
    IdError(i32),
    ImageError(i32),
    LanguageCodeError(i32),
    MediaBundleError(i32),
    MediaUploadError(i32),
    MediaFileError(i32),
    MerchantCenterError(i32),
    MultiplierError(i32),
    NewResourceCreationError(i32),
    NotEmptyError(i32),
    NullError(i32),
    OperatorError(i32),
    RangeError(i32),
    RecommendationError(i32),
    RecommendationSubscriptionError(i32),
    RegionCodeError(i32),
    SettingError(i32),
    StringFormatError(i32),
    StringLengthError(i32),
    OperationAccessDeniedError(i32),
    ResourceAccessDeniedError(i32),
    ResourceCountLimitExceededError(i32),
    YoutubeVideoRegistrationError(i32),
    AdGroupBidModifierError(i32),
    ContextError(i32),
    FieldError(i32),
    SharedSetError(i32),
    SharedCriterionError(i32),
    CampaignSharedSetError(i32),
    ConversionActionError(i32),
    ConversionAdjustmentUploadError(i32),
    ConversionCustomVariableError(i32),
    ConversionUploadError(i32),
    ConversionValueRuleError(i32),
    ConversionValueRuleSetError(i32),
    HeaderError(i32),
    DatabaseError(i32),
    PolicyFindingError(i32),
    EnumError(i32),
    KeywordPlanError(i32),
    KeywordPlanCampaignError(i32),
    KeywordPlanCampaignKeywordError(i32),
    KeywordPlanAdGroupError(i32),
    KeywordPlanAdGroupKeywordError(i32),
    KeywordPlanIdeaError(i32),
    AccountBudgetProposalError(i32),
    UserListError(i32),
    ChangeEventError(i32),
    ChangeStatusError(i32),
    FeedError(i32),
    GeoTargetConstantSuggestionError(i32),
    CampaignDraftError(i32),
    FeedItemError(i32),
    LabelError(i32),
    BillingSetupError(i32),
    CustomerClientLinkError(i32),
    CustomerManagerLinkError(i32),
    FeedMappingError(i32),
    CustomerFeedError(i32),
    AdGroupFeedError(i32),
    CampaignFeedError(i32),
    CustomInterestError(i32),
    CampaignExperimentError(i32),
    ExtensionFeedItemError(i32),
    AdParameterError(i32),
    FeedItemValidationError(i32),
    ExtensionSettingError(i32),
    FeedItemSetError(i32),
    FeedItemSetLinkError(i32),
    FeedItemTargetError(i32),
    PolicyViolationError(i32),
    PartialFailureError(i32),
    PolicyValidationParameterError(i32),
    SizeLimitError(i32),
    OfflineUserDataJobError(i32),
    NotAllowlistedError(i32),
    ManagerLinkError(i32),
    CurrencyCodeError(i32),
    ExperimentError(i32),
    AccessInvitationError(i32),
    ReachPlanError(i32),
    InvoiceError(i32),
    PaymentsAccountError(i32),
    TimeZoneError(i32),
    AssetLinkError(i32),
    UserDataError(i32),
    BatchJobError(i32),
    AccountLinkError(i32),
    ThirdPartyAppAnalyticsLinkError(i32),
    CustomerUserAccessError(i32),
    CustomAudienceError(i32),
    AudienceError(i32),
    SearchTermInsightError(i32),
    SmartCampaignError(i32),
    ExperimentArmError(i32),
    AudienceInsightsError(i32),
    ProductLinkError(i32),
    CustomerSkAdNetworkConversionValueSchemaError(i32),
    CurrencyError(i32),
    AssetGroupSignalError(i32),
    ProductLinkInvitationError(i32),
    CustomerLifecycleGoalError(i32),
    CampaignLifecycleGoalError(i32),
    IdentityVerificationError(i32),
    UserListCustomerTypeError(i32),
    ShoppingProductError(i32),
    AutomaticallyCreatedAssetRemovalError(i32),
    ShareablePreviewError(i32),

}

The list of error enums

Variants

RequestError(i32)

An error caused by the request

BiddingStrategyError(i32)

An error with a Bidding Strategy mutate.

UrlFieldError(i32)

An error with a URL field mutate.

ListOperationError(i32)

An error with a list operation.

QueryError(i32)

An error with an AWQL query

MutateError(i32)

An error with a mutate

FieldMaskError(i32)

An error with a field mask

AuthorizationError(i32)

An error encountered when trying to authorize a user.

InternalError(i32)

An unexpected server-side error.

QuotaError(i32)

An error with the amount of quota remaining.

AdError(i32)

An error with an Ad Group Ad mutate.

AdGroupError(i32)

An error with an Ad Group mutate.

CampaignBudgetError(i32)

An error with a Campaign Budget mutate.

CampaignError(i32)

An error with a Campaign mutate.

VideoCampaignError(i32)

An error with a Video Campaign mutate.

AuthenticationError(i32)

Indicates failure to properly authenticate user.

AdGroupCriterionCustomizerError(i32)

The reasons for the ad group criterion customizer error.

AdGroupCriterionError(i32)

Indicates failure to properly authenticate user.

AdGroupCustomizerError(i32)

The reasons for the ad group customizer error.

AdCustomizerError(i32)

The reasons for the ad customizer error

AdGroupAdError(i32)

The reasons for the ad group ad error

AdSharingError(i32)

The reasons for the ad sharing error

AdxError(i32)

The reasons for the adx error

AssetError(i32)

The reasons for the asset error

AssetGroupAssetError(i32)

The reasons for the asset group asset error

AssetGroupListingGroupFilterError(i32)

The reasons for the asset group listing group filter error

AssetGroupError(i32)

The reasons for the asset group error

AssetSetAssetError(i32)

The reasons for the asset set asset error

AssetSetLinkError(i32)

The reasons for the asset set link error

AssetSetError(i32)

The reasons for the asset set error

BiddingError(i32)

The reasons for the bidding errors

CampaignCriterionError(i32)

The reasons for the campaign criterion error

CampaignConversionGoalError(i32)

The reasons for the campaign conversion goal error

CampaignCustomizerError(i32)

The reasons for the campaign customizer error.

CollectionSizeError(i32)

The reasons for the collection size error

ConversionGoalCampaignConfigError(i32)

The reasons for the conversion goal campaign config error

CountryCodeError(i32)

The reasons for the country code error

CriterionError(i32)

The reasons for the criterion error

CustomConversionGoalError(i32)

The reasons for the custom conversion goal error

CustomerCustomizerError(i32)

The reasons for the customer customizer error.

CustomerError(i32)

The reasons for the customer error

CustomizerAttributeError(i32)

The reasons for the customizer attribute error.

DateError(i32)

The reasons for the date error

DateRangeError(i32)

The reasons for the date range error

DistinctError(i32)

The reasons for the distinct error

FeedAttributeReferenceError(i32)

The reasons for the feed attribute reference error

FunctionError(i32)

The reasons for the function error

FunctionParsingError(i32)

The reasons for the function parsing error

IdError(i32)

The reasons for the id error

ImageError(i32)

The reasons for the image error

LanguageCodeError(i32)

The reasons for the language code error

MediaBundleError(i32)

The reasons for the media bundle error

MediaUploadError(i32)

The reasons for media uploading errors.

MediaFileError(i32)

The reasons for the media file error

MerchantCenterError(i32)

Container for enum describing possible merchant center errors.

MultiplierError(i32)

The reasons for the multiplier error

NewResourceCreationError(i32)

The reasons for the new resource creation error

NotEmptyError(i32)

The reasons for the not empty error

NullError(i32)

The reasons for the null error

OperatorError(i32)

The reasons for the operator error

RangeError(i32)

The reasons for the range error

RecommendationError(i32)

The reasons for error in applying a recommendation

RecommendationSubscriptionError(i32)

The reasons for the recommendation subscription error.

RegionCodeError(i32)

The reasons for the region code error

SettingError(i32)

The reasons for the setting error

StringFormatError(i32)

The reasons for the string format error

StringLengthError(i32)

The reasons for the string length error

OperationAccessDeniedError(i32)

The reasons for the operation access denied error

ResourceAccessDeniedError(i32)

The reasons for the resource access denied error

ResourceCountLimitExceededError(i32)

The reasons for the resource count limit exceeded error

YoutubeVideoRegistrationError(i32)

The reasons for YouTube video registration errors.

AdGroupBidModifierError(i32)

The reasons for the ad group bid modifier error

ContextError(i32)

The reasons for the context error

FieldError(i32)

The reasons for the field error

SharedSetError(i32)

The reasons for the shared set error

SharedCriterionError(i32)

The reasons for the shared criterion error

CampaignSharedSetError(i32)

The reasons for the campaign shared set error

ConversionActionError(i32)

The reasons for the conversion action error

ConversionAdjustmentUploadError(i32)

The reasons for the conversion adjustment upload error

ConversionCustomVariableError(i32)

The reasons for the conversion custom variable error

ConversionUploadError(i32)

The reasons for the conversion upload error

ConversionValueRuleError(i32)

The reasons for the conversion value rule error

ConversionValueRuleSetError(i32)

The reasons for the conversion value rule set error

HeaderError(i32)

The reasons for the header error.

DatabaseError(i32)

The reasons for the database error.

PolicyFindingError(i32)

The reasons for the policy finding error.

EnumError(i32)

The reason for enum error.

KeywordPlanError(i32)

The reason for keyword plan error.

KeywordPlanCampaignError(i32)

The reason for keyword plan campaign error.

KeywordPlanCampaignKeywordError(i32)

The reason for keyword plan campaign keyword error.

KeywordPlanAdGroupError(i32)

The reason for keyword plan ad group error.

KeywordPlanAdGroupKeywordError(i32)

The reason for keyword plan ad group keyword error.

KeywordPlanIdeaError(i32)

The reason for keyword idea error.

AccountBudgetProposalError(i32)

The reasons for account budget proposal errors.

UserListError(i32)

The reasons for the user list error

ChangeEventError(i32)

The reasons for the change event error

ChangeStatusError(i32)

The reasons for the change status error

FeedError(i32)

The reasons for the feed error

GeoTargetConstantSuggestionError(i32)

The reasons for the geo target constant suggestion error.

CampaignDraftError(i32)

The reasons for the campaign draft error

FeedItemError(i32)

The reasons for the feed item error

LabelError(i32)

The reason for the label error.

BillingSetupError(i32)

The reasons for the billing setup error

CustomerClientLinkError(i32)

The reasons for the customer client link error

CustomerManagerLinkError(i32)

The reasons for the customer manager link error

FeedMappingError(i32)

The reasons for the feed mapping error

CustomerFeedError(i32)

The reasons for the customer feed error

AdGroupFeedError(i32)

The reasons for the ad group feed error

CampaignFeedError(i32)

The reasons for the campaign feed error

CustomInterestError(i32)

The reasons for the custom interest error

CampaignExperimentError(i32)

The reasons for the campaign experiment error

ExtensionFeedItemError(i32)

The reasons for the extension feed item error

AdParameterError(i32)

The reasons for the ad parameter error

FeedItemValidationError(i32)

The reasons for the feed item validation error

ExtensionSettingError(i32)

The reasons for the extension setting error

FeedItemSetError(i32)

The reasons for the feed item set error

FeedItemSetLinkError(i32)

The reasons for the feed item set link error

FeedItemTargetError(i32)

The reasons for the feed item target error

PolicyViolationError(i32)

The reasons for the policy violation error

PartialFailureError(i32)

The reasons for the mutate job error

PolicyValidationParameterError(i32)

The reasons for the policy validation parameter error

SizeLimitError(i32)

The reasons for the size limit error

OfflineUserDataJobError(i32)

The reasons for the offline user data job error.

NotAllowlistedError(i32)

The reasons for the not allowlisted error

ManagerLinkError(i32)

The reasons for the manager link error

CurrencyCodeError(i32)

The reasons for the currency code error

ExperimentError(i32)

The reasons for the experiment error

AccessInvitationError(i32)

The reasons for the access invitation error

ReachPlanError(i32)

The reasons for the reach plan error

InvoiceError(i32)

The reasons for the invoice error

PaymentsAccountError(i32)

The reasons for errors in payments accounts service

TimeZoneError(i32)

The reasons for the time zone error

AssetLinkError(i32)

The reasons for the asset link error

UserDataError(i32)

The reasons for the user data error.

BatchJobError(i32)

The reasons for the batch job error

AccountLinkError(i32)

The reasons for the account link status change error

ThirdPartyAppAnalyticsLinkError(i32)

The reasons for the third party app analytics link mutate error

CustomerUserAccessError(i32)

The reasons for the customer user access mutate error

CustomAudienceError(i32)

The reasons for the custom audience error

AudienceError(i32)

The reasons for the audience error

SearchTermInsightError(i32)

The reasons for the Search term insight error

SmartCampaignError(i32)

The reasons for the Smart campaign error

ExperimentArmError(i32)

The reasons for the experiment arm error

AudienceInsightsError(i32)

The reasons for the Audience Insights error

ProductLinkError(i32)

The reasons for the product link error

CustomerSkAdNetworkConversionValueSchemaError(i32)

The reasons for the customer SK Ad network conversion value schema error

CurrencyError(i32)

The reasons for the currency errors.

AssetGroupSignalError(i32)

The reasons for the asset group hint error

ProductLinkInvitationError(i32)

The reasons for the product link invitation error

CustomerLifecycleGoalError(i32)

The reasons for the customer lifecycle goal error

CampaignLifecycleGoalError(i32)

The reasons for the campaign lifecycle goal error

IdentityVerificationError(i32)

The reasons for an identity verification error.

UserListCustomerTypeError(i32)

The reasons for a user list customer type error.

ShoppingProductError(i32)

The reasons for error in querying shopping product.

AutomaticallyCreatedAssetRemovalError(i32)

The reasons for error in automatically created asset removal action.

ShareablePreviewError(i32)

The reasons for the shareable preview error.

Implementations

impl ErrorCode

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<ErrorCode>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.