Enum google_api_proto::google::ads::googleads::v17::errors::url_field_error_enum::UrlFieldError

#[repr(i32)]
pub enum UrlFieldError {

    Unspecified = 0,
    Unknown = 1,
    InvalidTrackingUrlTemplate = 2,
    InvalidTagInTrackingUrlTemplate = 3,
    MissingTrackingUrlTemplateTag = 4,
    MissingProtocolInTrackingUrlTemplate = 5,
    InvalidProtocolInTrackingUrlTemplate = 6,
    MalformedTrackingUrlTemplate = 7,
    MissingHostInTrackingUrlTemplate = 8,
    InvalidTldInTrackingUrlTemplate = 9,
    RedundantNestedTrackingUrlTemplateTag = 10,
    InvalidFinalUrl = 11,
    InvalidTagInFinalUrl = 12,
    RedundantNestedFinalUrlTag = 13,
    MissingProtocolInFinalUrl = 14,
    InvalidProtocolInFinalUrl = 15,
    MalformedFinalUrl = 16,
    MissingHostInFinalUrl = 17,
    InvalidTldInFinalUrl = 18,
    InvalidFinalMobileUrl = 19,
    InvalidTagInFinalMobileUrl = 20,
    RedundantNestedFinalMobileUrlTag = 21,
    MissingProtocolInFinalMobileUrl = 22,
    InvalidProtocolInFinalMobileUrl = 23,
    MalformedFinalMobileUrl = 24,
    MissingHostInFinalMobileUrl = 25,
    InvalidTldInFinalMobileUrl = 26,
    InvalidFinalAppUrl = 27,
    InvalidTagInFinalAppUrl = 28,
    RedundantNestedFinalAppUrlTag = 29,
    MultipleAppUrlsForOstype = 30,
    InvalidOstype = 31,
    InvalidProtocolForAppUrl = 32,
    InvalidPackageIdForAppUrl = 33,
    UrlCustomParametersCountExceedsLimit = 34,
    InvalidCharactersInUrlCustomParameterKey = 39,
    InvalidCharactersInUrlCustomParameterValue = 40,
    InvalidTagInUrlCustomParameterValue = 41,
    RedundantNestedUrlCustomParameterTag = 42,
    MissingProtocol = 43,
    InvalidProtocol = 52,
    InvalidUrl = 44,
    DestinationUrlDeprecated = 45,
    InvalidTagInUrl = 46,
    MissingUrlTag = 47,
    DuplicateUrlId = 48,
    InvalidUrlId = 49,
    FinalUrlSuffixMalformed = 50,
    InvalidTagInFinalUrlSuffix = 51,
    InvalidTopLevelDomain = 53,
    MalformedTopLevelDomain = 54,
    MalformedUrl = 55,
    MissingHost = 56,
    NullCustomParameterValue = 57,
    ValueTrackParameterNotSupported = 58,

}

Enum describing possible url field errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

InvalidTrackingUrlTemplate = 2

The tracking url template is invalid.

InvalidTagInTrackingUrlTemplate = 3

The tracking url template contains invalid tag.

MissingTrackingUrlTemplateTag = 4

The tracking url template must contain at least one tag (for example, {lpurl}), This applies only to tracking url template associated with website ads or product ads.

MissingProtocolInTrackingUrlTemplate = 5

The tracking url template must start with a valid protocol (or lpurl tag).

InvalidProtocolInTrackingUrlTemplate = 6

The tracking url template starts with an invalid protocol.

MalformedTrackingUrlTemplate = 7

The tracking url template contains illegal characters.

MissingHostInTrackingUrlTemplate = 8

The tracking url template must contain a host name (or lpurl tag).

InvalidTldInTrackingUrlTemplate = 9

The tracking url template has an invalid or missing top level domain extension.

RedundantNestedTrackingUrlTemplateTag = 10

The tracking url template contains nested occurrences of the same conditional tag (for example, {ifmobile:{ifmobile:x}}).

InvalidFinalUrl = 11

The final url is invalid.

InvalidTagInFinalUrl = 12

The final url contains invalid tag.

RedundantNestedFinalUrlTag = 13

The final url contains nested occurrences of the same conditional tag (for example, {ifmobile:{ifmobile:x}}).

MissingProtocolInFinalUrl = 14

The final url must start with a valid protocol.

InvalidProtocolInFinalUrl = 15

The final url starts with an invalid protocol.

MalformedFinalUrl = 16

The final url contains illegal characters.

MissingHostInFinalUrl = 17

The final url must contain a host name.

InvalidTldInFinalUrl = 18

The tracking url template has an invalid or missing top level domain extension.

InvalidFinalMobileUrl = 19

The final mobile url is invalid.

InvalidTagInFinalMobileUrl = 20

The final mobile url contains invalid tag.

RedundantNestedFinalMobileUrlTag = 21

The final mobile url contains nested occurrences of the same conditional tag (for example, {ifmobile:{ifmobile:x}}).

MissingProtocolInFinalMobileUrl = 22

The final mobile url must start with a valid protocol.

InvalidProtocolInFinalMobileUrl = 23

The final mobile url starts with an invalid protocol.

MalformedFinalMobileUrl = 24

The final mobile url contains illegal characters.

MissingHostInFinalMobileUrl = 25

The final mobile url must contain a host name.

InvalidTldInFinalMobileUrl = 26

The tracking url template has an invalid or missing top level domain extension.

InvalidFinalAppUrl = 27

The final app url is invalid.

InvalidTagInFinalAppUrl = 28

The final app url contains invalid tag.

RedundantNestedFinalAppUrlTag = 29

The final app url contains nested occurrences of the same conditional tag (for example, {ifmobile:{ifmobile:x}}).

MultipleAppUrlsForOstype = 30

More than one app url found for the same OS type.

InvalidOstype = 31

The OS type given for an app url is not valid.

InvalidProtocolForAppUrl = 32

The protocol given for an app url is not valid. (For example, “android-app://”)

InvalidPackageIdForAppUrl = 33

The package id (app id) given for an app url is not valid.

UrlCustomParametersCountExceedsLimit = 34

The number of url custom parameters for an resource exceeds the maximum limit allowed.

InvalidCharactersInUrlCustomParameterKey = 39

An invalid character appears in the parameter key.

InvalidCharactersInUrlCustomParameterValue = 40

An invalid character appears in the parameter value.

InvalidTagInUrlCustomParameterValue = 41

The url custom parameter value fails url tag validation.

RedundantNestedUrlCustomParameterTag = 42

The custom parameter contains nested occurrences of the same conditional tag (for example, {ifmobile:{ifmobile:x}}).

MissingProtocol = 43

The protocol (http:// or https://) is missing.

InvalidProtocol = 52

Unsupported protocol in URL. Only http and https are supported.

InvalidUrl = 44

The url is invalid.

DestinationUrlDeprecated = 45

Destination Url is deprecated.

InvalidTagInUrl = 46

The url contains invalid tag.

MissingUrlTag = 47

The url must contain at least one tag (for example, {lpurl}).

DuplicateUrlId = 48

Duplicate url id.

InvalidUrlId = 49

Invalid url id.

FinalUrlSuffixMalformed = 50

The final url suffix cannot begin with ‘?’ or ‘&’ characters and must be a valid query string.

InvalidTagInFinalUrlSuffix = 51

The final url suffix cannot contain {lpurl} related or {ignore} tags.

InvalidTopLevelDomain = 53

The top level domain is invalid, for example, not a public top level domain listed in publicsuffix.org.

MalformedTopLevelDomain = 54

Malformed top level domain in URL.

MalformedUrl = 55

Malformed URL.

MissingHost = 56

No host found in URL.

NullCustomParameterValue = 57

Custom parameter value cannot be null.

ValueTrackParameterNotSupported = 58

Track parameter is not supported.

Implementations

impl UrlFieldError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of UrlFieldError.

pub fn from_i32(value: i32) -> Option<UrlFieldError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a UrlFieldError, or None if value is not a valid variant.

impl UrlFieldError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.