Enum google_api_proto::google::ads::googleads::v17::errors::user_list_error_enum::UserListError

#[repr(i32)]
pub enum UserListError {

    Unspecified = 0,
    Unknown = 1,
    ExternalRemarketingUserListMutateNotSupported = 2,
    ConcreteTypeRequired = 3,
    ConversionTypeIdRequired = 4,
    DuplicateConversionTypes = 5,
    InvalidConversionType = 6,
    InvalidDescription = 7,
    InvalidName = 8,
    InvalidType = 9,
    CanNotAddLogicalListAsLogicalListOperand = 10,
    InvalidUserListLogicalRuleOperand = 11,
    NameAlreadyUsed = 12,
    NewConversionTypeNameRequired = 13,
    ConversionTypeNameAlreadyUsed = 14,
    OwnershipRequiredForSet = 15,
    UserListMutateNotSupported = 16,
    InvalidRule = 17,
    InvalidDateRange = 27,
    CanNotMutateSensitiveUserlist = 28,
    MaxNumRulebasedUserlists = 29,
    CannotModifyBillableRecordCount = 30,
    AppIdNotSet = 31,
    UserlistNameIsReservedForSystemList = 32,
    AdvertiserNotOnAllowlistForUsingUploadedData = 37,
    RuleTypeIsNotSupported = 34,
    CanNotAddASimilarUserlistAsLogicalListOperand = 35,
    CanNotMixCrmBasedInLogicalListWithOtherLists = 36,
    AppIdNotAllowed = 39,
    CannotMutateSystemList = 40,
    MobileAppIsSensitive = 41,
    SeedListDoesNotExist = 42,
    InvalidSeedListAccessReason = 43,
    InvalidSeedListType = 44,
    InvalidCountryCodes = 45,

}

Enum describing possible user list errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

ExternalRemarketingUserListMutateNotSupported = 2

Creating and updating external remarketing user lists is not supported.

ConcreteTypeRequired = 3

Concrete type of user list is required.

ConversionTypeIdRequired = 4

Creating/updating user list conversion types requires specifying the conversion type Id.

DuplicateConversionTypes = 5

Remarketing user list cannot have duplicate conversion types.

InvalidConversionType = 6

Conversion type is invalid/unknown.

InvalidDescription = 7

User list description is empty or invalid.

InvalidName = 8

User list name is empty or invalid.

InvalidType = 9

Type of the UserList does not match.

CanNotAddLogicalListAsLogicalListOperand = 10

Embedded logical user lists are not allowed.

InvalidUserListLogicalRuleOperand = 11

User list rule operand is invalid.

NameAlreadyUsed = 12

Name is already being used for another user list for the account.

NewConversionTypeNameRequired = 13

Name is required when creating a new conversion type.

ConversionTypeNameAlreadyUsed = 14

The given conversion type name has been used.

OwnershipRequiredForSet = 15

Only an owner account may edit a user list.

UserListMutateNotSupported = 16

Creating user list without setting type in oneof user_list field, or creating/updating read-only user list types is not allowed.

InvalidRule = 17

Rule is invalid.

InvalidDateRange = 27

The specified date range is empty.

CanNotMutateSensitiveUserlist = 28

A UserList which is privacy sensitive or legal rejected cannot be mutated by external users.

MaxNumRulebasedUserlists = 29

Maximum number of rulebased user lists a customer can have.

CannotModifyBillableRecordCount = 30

BasicUserList’s billable record field cannot be modified once it is set.

AppIdNotSet = 31

crm_based_user_list.app_id field must be set when upload_key_type is MOBILE_ADVERTISING_ID.

UserlistNameIsReservedForSystemList = 32

Name of the user list is reserved for system generated lists and cannot be used.

AdvertiserNotOnAllowlistForUsingUploadedData = 37

Advertiser needs to be on the allow-list to use remarketing lists created from advertiser uploaded data (for example, Customer Match lists).

RuleTypeIsNotSupported = 34

The provided rule_type is not supported for the user list.

CanNotAddASimilarUserlistAsLogicalListOperand = 35

Similar user list cannot be used as a logical user list operand.

CanNotMixCrmBasedInLogicalListWithOtherLists = 36

Logical user list should not have a mix of CRM based user list and other types of lists in its rules.

AppIdNotAllowed = 39

crm_based_user_list.app_id field can only be set when upload_key_type is MOBILE_ADVERTISING_ID.

CannotMutateSystemList = 40

Google system generated user lists cannot be mutated.

MobileAppIsSensitive = 41

The mobile app associated with the remarketing list is sensitive.

SeedListDoesNotExist = 42

One or more given seed lists do not exist.

InvalidSeedListAccessReason = 43

One or more given seed lists are not accessible to the current user.

InvalidSeedListType = 44

One or more given seed lists have an unsupported type.

InvalidCountryCodes = 45

One or more invalid country codes are added to Lookalike UserList.

Implementations

impl UserListError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of UserListError.

pub fn from_i32(value: i32) -> Option<UserListError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a UserListError, or None if value is not a valid variant.

impl UserListError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.