Module google_api_proto::google::ads::googleads::v17::resources::campaign

Nested message and enum types in Campaign.

Modules

video_campaign_settings

Nested message and enum types in VideoCampaignSettings.

Structs

AppCampaignSetting

Campaign-level settings for App Campaigns.

AssetAutomationSetting

Asset automation setting contains pair of AssetAutomationType and the asset automation opt-in/out status

AudienceSetting

Settings for the audience targeting.

CategoryBid

Category bids in LocalServicesReportingCampaignSettings.

DemandGenCampaignSettings

Settings for Demand Gen campaign.

DynamicSearchAdsSetting

The setting for controlling Dynamic Search Ads (DSA).

GeoTargetTypeSetting

Represents a collection of settings related to ads geotargeting.

HotelSettingInfo

Campaign-level settings for hotel ads.

LocalCampaignSetting

Campaign setting for local campaigns.

LocalServicesCampaignSettings

Settings for LocalServicesCampaign subresource.

NetworkSettings

The network settings for the campaign.

OptimizationGoalSetting

Optimization goal setting for this campaign, which includes a set of optimization goal types.

PerformanceMaxUpgrade

Information about a campaign being upgraded to Performance Max.

SelectiveOptimization

Selective optimization setting for this campaign, which includes a set of conversion actions to optimize this campaign towards. This feature only applies to app campaigns that use MULTI_CHANNEL as AdvertisingChannelType and APP_CAMPAIGN or APP_CAMPAIGN_FOR_ENGAGEMENT as AdvertisingChannelSubType.

ShoppingSetting

The setting for Shopping campaigns. Defines the universe of products that can be advertised by the campaign, and how this campaign interacts with other Shopping campaigns.

TrackingSetting

Campaign-level settings for tracking information.

TravelCampaignSettings

Settings for Travel campaign.

VanityPharma

Describes how unbranded pharma ads will be displayed.

VideoCampaignSettings

Settings for Video campaign.

Enums

CampaignBiddingStrategy

The bidding strategy for the campaign.