Struct google_api_proto::google::ads::googleads::v17::resources::ConversionAction

pub struct ConversionAction {

    pub resource_name: String,
    pub id: Option<i64>,
    pub name: Option<String>,
    pub status: i32,
    pub type: i32,
    pub origin: i32,
    pub primary_for_goal: Option<bool>,
    pub category: i32,
    pub owner_customer: Option<String>,
    pub include_in_conversions_metric: Option<bool>,
    pub click_through_lookback_window_days: Option<i64>,
    pub view_through_lookback_window_days: Option<i64>,
    pub value_settings: Option<ValueSettings>,
    pub counting_type: i32,
    pub attribution_model_settings: Option<AttributionModelSettings>,
    pub tag_snippets: Vec<TagSnippet>,
    pub phone_call_duration_seconds: Option<i64>,
    pub app_id: Option<String>,
    pub mobile_app_vendor: i32,
    pub firebase_settings: Option<FirebaseSettings>,
    pub third_party_app_analytics_settings: Option<ThirdPartyAppAnalyticsSettings>,
    pub google_analytics_4_settings: Option<GoogleAnalytics4Settings>,

}

A conversion action.

Fields
resource_name: String

Immutable. The resource name of the conversion action. Conversion action resource names have the form:

customers/{customer_id}/conversionActions/{conversion_action_id}

id: Option<i64>

Output only. The ID of the conversion action.

name: Option<String>

The name of the conversion action.

This field is required and should not be empty when creating new conversion actions.

status: i32

The status of this conversion action for conversion event accrual.

type: i32

Immutable. The type of this conversion action.

origin: i32

Output only. The conversion origin of this conversion action.

primary_for_goal: Option<bool>

If a conversion action’s primary_for_goal bit is false, the conversion action is non-biddable for all campaigns regardless of their customer conversion goal or campaign conversion goal. However, custom conversion goals do not respect primary_for_goal, so if a campaign has a custom conversion goal configured with a primary_for_goal = false conversion action, that conversion action is still biddable. By default, primary_for_goal will be true if not set. In V9, primary_for_goal can only be set to false after creation through an ‘update’ operation because it’s not declared as optional.

category: i32

The category of conversions reported for this conversion action.

owner_customer: Option<String>

Output only. The resource name of the conversion action owner customer, or null if this is a system-defined conversion action.

include_in_conversions_metric: Option<bool>

Whether this conversion action should be included in the “conversions” metric.

click_through_lookback_window_days: Option<i64>

The maximum number of days that may elapse between an interaction (for example, a click) and a conversion event.

view_through_lookback_window_days: Option<i64>

The maximum number of days which may elapse between an impression and a conversion without an interaction.

value_settings: Option<ValueSettings>

Settings related to the value for conversion events associated with this conversion action.

counting_type: i32

How to count conversion events for the conversion action.

attribution_model_settings: Option<AttributionModelSettings>

Settings related to this conversion action’s attribution model.

tag_snippets: Vec<TagSnippet>

Output only. The snippets used for tracking conversions.

phone_call_duration_seconds: Option<i64>

The phone call duration in seconds after which a conversion should be reported for this conversion action.

The value must be between 0 and 10000, inclusive.

app_id: Option<String>

App ID for an app conversion action.

mobile_app_vendor: i32

Output only. Mobile app vendor for an app conversion action.

firebase_settings: Option<FirebaseSettings>

Output only. Firebase settings for Firebase conversion types.

third_party_app_analytics_settings: Option<ThirdPartyAppAnalyticsSettings>

Output only. Third Party App Analytics settings for third party conversion types.

google_analytics_4_settings: Option<GoogleAnalytics4Settings>

Output only. Google Analytics 4 settings for Google Analytics 4 conversion types.

Implementations

impl ConversionAction

pub fn status(&self) -> ConversionActionStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: ConversionActionStatus)

Sets status to the provided enum value.

pub fn type(&self) -> ConversionActionType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: ConversionActionType)

Sets type to the provided enum value.

pub fn category(&self) -> ConversionActionCategory

Returns the enum value of category, or the default if the field is set to an invalid enum value.

pub fn set_category(&mut self, value: ConversionActionCategory)

Sets category to the provided enum value.

pub fn counting_type(&self) -> ConversionActionCountingType

Returns the enum value of counting_type, or the default if the field is set to an invalid enum value.

pub fn set_counting_type(&mut self, value: ConversionActionCountingType)

Sets counting_type to the provided enum value.

pub fn mobile_app_vendor(&self) -> MobileAppVendor

Returns the enum value of mobile_app_vendor, or the default if the field is set to an invalid enum value.

pub fn set_mobile_app_vendor(&mut self, value: MobileAppVendor)

Sets mobile_app_vendor to the provided enum value.

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn owner_customer(&self) -> &str

Returns the value of owner_customer, or the default value if owner_customer is unset.

pub fn include_in_conversions_metric(&self) -> bool

Returns the value of include_in_conversions_metric, or the default value if include_in_conversions_metric is unset.

pub fn click_through_lookback_window_days(&self) -> i64

Returns the value of click_through_lookback_window_days, or the default value if click_through_lookback_window_days is unset.

pub fn view_through_lookback_window_days(&self) -> i64

Returns the value of view_through_lookback_window_days, or the default value if view_through_lookback_window_days is unset.

pub fn phone_call_duration_seconds(&self) -> i64

Returns the value of phone_call_duration_seconds, or the default value if phone_call_duration_seconds is unset.

pub fn app_id(&self) -> &str

Returns the value of app_id, or the default value if app_id is unset.

pub fn origin(&self) -> ConversionOrigin

Returns the enum value of origin, or the default if the field is set to an invalid enum value.

pub fn set_origin(&mut self, value: ConversionOrigin)

Sets origin to the provided enum value.

pub fn primary_for_goal(&self) -> bool

Returns the value of primary_for_goal, or the default value if primary_for_goal is unset.