Struct google_api_proto::google::ads::googleads::v17::resources::LocalServicesEmployee

pub struct LocalServicesEmployee {

    pub resource_name: String,
    pub id: Option<i64>,
    pub creation_date_time: String,
    pub status: i32,
    pub type: i32,
    pub university_degrees: Vec<UniversityDegree>,
    pub residencies: Vec<Residency>,
    pub fellowships: Vec<Fellowship>,
    pub job_title: Option<String>,
    pub year_started_practicing: Option<i32>,
    pub languages_spoken: Vec<String>,
    pub category_ids: Vec<String>,
    pub national_provider_id_number: Option<String>,
    pub email_address: Option<String>,
    pub first_name: Option<String>,
    pub middle_name: Option<String>,
    pub last_name: Option<String>,

}

A local services employee resource.

Fields
resource_name: String

Immutable. The resource name of the Local Services Verification. Local Services Verification resource names have the form:

customers/{customer_id}/localServicesEmployees/{gls_employee_id}

id: Option<i64>

Output only. The ID of the employee.

creation_date_time: String

Output only. Timestamp of employee creation. The format is “YYYY-MM-DD HH:MM:SS” in the Google Ads account’s timezone. Examples: “2018-03-05 09:15:00” or “2018-02-01 14:34:30”

status: i32

Output only. Employee status, such as DELETED or ENABLED.

type: i32

Output only. Employee type.

university_degrees: Vec<UniversityDegree>

Output only. A list of degrees this employee has obtained, and wants to feature.

residencies: Vec<Residency>

Output only. The institutions where the employee has completed their residency.

fellowships: Vec<Fellowship>

Output only. The institutions where the employee has completed their fellowship.

job_title: Option<String>

Output only. Job title for this employee, such as “Senior partner” in legal verticals.

year_started_practicing: Option<i32>

Output only. The year that this employee started practicing in this field.

languages_spoken: Vec<String>

Output only. Languages that the employee speaks, represented as language tags from https://developers.google.com/admin-sdk/directory/v1/languages

category_ids: Vec<String>

Output only. Category of the employee. A list of Local Services category IDs can be found at https://developers.google.com/google-ads/api/data/codes-formats#local_services_ids.

national_provider_id_number: Option<String>

Output only. NPI id associated with the employee.

email_address: Option<String>

Output only. Email address of the employee.

first_name: Option<String>

Output only. First name of the employee.

middle_name: Option<String>

Output only. Middle name of the employee.

last_name: Option<String>

Output only. Last name of the employee.

Implementations

impl LocalServicesEmployee

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn status(&self) -> LocalServicesEmployeeStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: LocalServicesEmployeeStatus)

Sets status to the provided enum value.

pub fn type(&self) -> LocalServicesEmployeeType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: LocalServicesEmployeeType)

Sets type to the provided enum value.

pub fn job_title(&self) -> &str

Returns the value of job_title, or the default value if job_title is unset.

pub fn year_started_practicing(&self) -> i32

Returns the value of year_started_practicing, or the default value if year_started_practicing is unset.

pub fn national_provider_id_number(&self) -> &str

Returns the value of national_provider_id_number, or the default value if national_provider_id_number is unset.

pub fn email_address(&self) -> &str

Returns the value of email_address, or the default value if email_address is unset.

pub fn first_name(&self) -> &str

Returns the value of first_name, or the default value if first_name is unset.

pub fn middle_name(&self) -> &str

Returns the value of middle_name, or the default value if middle_name is unset.

pub fn last_name(&self) -> &str

Returns the value of last_name, or the default value if last_name is unset.