Struct google_api_proto::google::ads::googleads::v17::resources::ShoppingProduct

pub struct ShoppingProduct {

    pub resource_name: String,
    pub merchant_center_id: i64,
    pub channel: i32,
    pub language_code: String,
    pub feed_label: String,
    pub item_id: String,
    pub multi_client_account_id: Option<i64>,
    pub title: Option<String>,
    pub brand: Option<String>,
    pub price_micros: Option<i64>,
    pub currency_code: Option<String>,
    pub channel_exclusivity: Option<i32>,
    pub condition: Option<i32>,
    pub availability: Option<i32>,
    pub target_countries: Vec<String>,
    pub custom_attribute0: Option<String>,
    pub custom_attribute1: Option<String>,
    pub custom_attribute2: Option<String>,
    pub custom_attribute3: Option<String>,
    pub custom_attribute4: Option<String>,
    pub category_level1: Option<String>,
    pub category_level2: Option<String>,
    pub category_level3: Option<String>,
    pub category_level4: Option<String>,
    pub category_level5: Option<String>,
    pub product_type_level1: Option<String>,
    pub product_type_level2: Option<String>,
    pub product_type_level3: Option<String>,
    pub product_type_level4: Option<String>,
    pub product_type_level5: Option<String>,
    pub effective_max_cpc_micros: Option<i64>,
    pub status: i32,
    pub issues: Vec<ProductIssue>,
    pub campaign: Option<String>,
    pub ad_group: Option<String>,

}

A shopping product from Google Merchant Center that can be advertised by campaigns.

The resource returns currently existing products from Google Merchant Center accounts linked with the customer. A campaign includes a product by specifying its merchant id (or, if available, the Multi Client Account id of the merchant) in the ShoppingSetting, and can limit the inclusion to products having a specified feed label. Standard Shopping campaigns can also limit the inclusion through a campaign_criterion.listing_scope.

Queries to this resource specify a scope: Account:

Filters on campaigns or ad groups are not specified.

All products from the linked Google Merchant Center accounts are returned.

Metrics and some fields (see the per-field documentation) are aggregated across all Shopping and Performance Max campaigns that include a product. Campaign:

An equality filter on campaign is specified. Supported campaign types are Shopping, Performance Max, Demand Gen, Video.

Only products that are included by the specified campaign are returned.

Metrics and some fields (see the per-field documentation) are restricted to the specified campaign. Ad group:

An equality filter on ad group and campaign is specified. Supported campaign types are Shopping, Demand Gen, Video.

Only products that are included by the specified campaign are returned.

Metrics and some fields (see the per-field documentation) are restricted to the specified ad group. Note that segmentation by date segments is not permitted and will return UNSUPPORTED_DATE_SEGMENTATION error. On the other hand, filtering on date segments is allowed.

Fields
resource_name: String

Output only. The resource name of the shopping product. Shopping product resource names have the form:

customers/{customer_id}/shoppingProducts/{merchant_center_id}~{channel}~{language_code}~{feed_label}~{item_id}

merchant_center_id: i64

Output only. The id of the merchant that owns the product.

channel: i32

Output only. The product channel describing the locality of the product.

language_code: String

Output only. The language code as provided by the merchant, in BCP 47 format.

feed_label: String

Output only. The product feed label as provided by the merchant.

item_id: String

Output only. The item id of the product as provided by the merchant.

multi_client_account_id: Option<i64>

Output only. The id of the Multi Client Account of the merchant, if present.

title: Option<String>

Output only. The title of the product as provided by the merchant.

brand: Option<String>

Output only. The brand of the product as provided by the merchant.

price_micros: Option<i64>

Output only. The price of the product in micros as provided by the merchant, in the currency specified in currency_code (e.g. $2.97 is reported as 2970000).

currency_code: Option<String>

Output only. The currency code as provided by the merchant, in ISO 4217 format.

channel_exclusivity: Option<i32>

Output only. The channel exclusivity of the product as provided by the merchant.

condition: Option<i32>

Output only. The condition of the product as provided by the merchant.

availability: Option<i32>

Output only. The availability of the product as provided by the merchant.

target_countries: Vec<String>

Output only. Upper-case two-letter ISO 3166-1 code of the regions where the product is intended to be shown in ads.

custom_attribute0: Option<String>

Output only. The custom attribute 0 of the product as provided by the merchant.

custom_attribute1: Option<String>

Output only. The custom attribute 1 of the product as provided by the merchant.

custom_attribute2: Option<String>

Output only. The custom attribute 2 of the product as provided by the merchant.

custom_attribute3: Option<String>

Output only. The custom attribute 3 of the product as provided by the merchant.

custom_attribute4: Option<String>

Output only. The custom attribute 4 of the product as provided by the merchant.

category_level1: Option<String>

Output only. The category level 1 of the product.

category_level2: Option<String>

Output only. The category level 2 of the product.

category_level3: Option<String>

Output only. The category level 3 of the product.

category_level4: Option<String>

Output only. The category level 4 of the product.

category_level5: Option<String>

Output only. The category level 5 of the product.

product_type_level1: Option<String>

Output only. The product type level 1 as provided by the merchant.

product_type_level2: Option<String>

Output only. The product type level 2 as provided by the merchant.

product_type_level3: Option<String>

Output only. The product type level 3 as provided by the merchant.

product_type_level4: Option<String>

Output only. The product type level 4 as provided by the merchant.

product_type_level5: Option<String>

Output only. The product type level 5 as provided by the merchant.

effective_max_cpc_micros: Option<i64>

Output only. The effective maximum cost-per-click (effective max. CPC) of the product. This field is available only if the query specifies the campaign or ad group scope, and if the campaign uses manual bidding. The value is the highest bid set for the product in product groups across all enabled ad groups. It represents the most you’re willing to pay for a click on the product. This field can take up to 24 hours to update.

status: i32

Output only. The status that indicates whether the product can show in ads. The value of this field is restricted to the scope specified in the query, see the documentation of the resource. This field can take up to 24 hours to update.

issues: Vec<ProductIssue>

Output only. The list of issues affecting whether the product can show in ads. The value of this field is restricted to the scope specified in the query, see the documentation of the resource. This field can take up to 24 hours to update.

campaign: Option<String>

Output only. A campaign that includes the product. This field is selectable only in the campaign scope, which requires an equality filter on campaign.

ad_group: Option<String>

Output only. An ad group of a campaign that includes the product. This field is selectable only in the ad group scope, which requires an equality filter on campaign and ad_group.

Implementations

impl ShoppingProduct

pub fn channel(&self) -> ProductChannel

Returns the enum value of channel, or the default if the field is set to an invalid enum value.

pub fn set_channel(&mut self, value: ProductChannel)

Sets channel to the provided enum value.

pub fn multi_client_account_id(&self) -> i64

Returns the value of multi_client_account_id, or the default value if multi_client_account_id is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn price_micros(&self) -> i64

Returns the value of price_micros, or the default value if price_micros is unset.

pub fn currency_code(&self) -> &str

Returns the value of currency_code, or the default value if currency_code is unset.

pub fn channel_exclusivity(&self) -> ProductChannelExclusivity

Returns the enum value of channel_exclusivity, or the default if the field is unset or set to an invalid enum value.

pub fn set_channel_exclusivity(&mut self, value: ProductChannelExclusivity)

Sets channel_exclusivity to the provided enum value.

pub fn condition(&self) -> ProductCondition

Returns the enum value of condition, or the default if the field is unset or set to an invalid enum value.

pub fn set_condition(&mut self, value: ProductCondition)

Sets condition to the provided enum value.

pub fn availability(&self) -> ProductAvailability

Returns the enum value of availability, or the default if the field is unset or set to an invalid enum value.

pub fn set_availability(&mut self, value: ProductAvailability)

Sets availability to the provided enum value.

pub fn custom_attribute0(&self) -> &str

Returns the value of custom_attribute0, or the default value if custom_attribute0 is unset.

pub fn custom_attribute1(&self) -> &str

Returns the value of custom_attribute1, or the default value if custom_attribute1 is unset.

pub fn custom_attribute2(&self) -> &str

Returns the value of custom_attribute2, or the default value if custom_attribute2 is unset.

pub fn custom_attribute3(&self) -> &str

Returns the value of custom_attribute3, or the default value if custom_attribute3 is unset.

pub fn custom_attribute4(&self) -> &str

Returns the value of custom_attribute4, or the default value if custom_attribute4 is unset.

pub fn category_level1(&self) -> &str

Returns the value of category_level1, or the default value if category_level1 is unset.

pub fn category_level2(&self) -> &str

Returns the value of category_level2, or the default value if category_level2 is unset.

pub fn category_level3(&self) -> &str

Returns the value of category_level3, or the default value if category_level3 is unset.

pub fn category_level4(&self) -> &str

Returns the value of category_level4, or the default value if category_level4 is unset.

pub fn category_level5(&self) -> &str

Returns the value of category_level5, or the default value if category_level5 is unset.

pub fn product_type_level1(&self) -> &str

Returns the value of product_type_level1, or the default value if product_type_level1 is unset.

pub fn product_type_level2(&self) -> &str

Returns the value of product_type_level2, or the default value if product_type_level2 is unset.

pub fn product_type_level3(&self) -> &str

Returns the value of product_type_level3, or the default value if product_type_level3 is unset.

pub fn product_type_level4(&self) -> &str

Returns the value of product_type_level4, or the default value if product_type_level4 is unset.

pub fn product_type_level5(&self) -> &str

Returns the value of product_type_level5, or the default value if product_type_level5 is unset.

pub fn effective_max_cpc_micros(&self) -> i64

Returns the value of effective_max_cpc_micros, or the default value if effective_max_cpc_micros is unset.

pub fn status(&self) -> ProductStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: ProductStatus)

Sets status to the provided enum value.

pub fn campaign(&self) -> &str

Returns the value of campaign, or the default value if campaign is unset.

pub fn ad_group(&self) -> &str

Returns the value of ad_group, or the default value if ad_group is unset.