Enum google_api_proto::google::ads::googleads::v7::errors::url_field_error_enum::UrlFieldError

#[repr(i32)]
pub enum UrlFieldError {

    Unspecified,
    Unknown,
    InvalidTrackingUrlTemplate,
    InvalidTagInTrackingUrlTemplate,
    MissingTrackingUrlTemplateTag,
    MissingProtocolInTrackingUrlTemplate,
    InvalidProtocolInTrackingUrlTemplate,
    MalformedTrackingUrlTemplate,
    MissingHostInTrackingUrlTemplate,
    InvalidTldInTrackingUrlTemplate,
    RedundantNestedTrackingUrlTemplateTag,
    InvalidFinalUrl,
    InvalidTagInFinalUrl,
    RedundantNestedFinalUrlTag,
    MissingProtocolInFinalUrl,
    InvalidProtocolInFinalUrl,
    MalformedFinalUrl,
    MissingHostInFinalUrl,
    InvalidTldInFinalUrl,
    InvalidFinalMobileUrl,
    InvalidTagInFinalMobileUrl,
    RedundantNestedFinalMobileUrlTag,
    MissingProtocolInFinalMobileUrl,
    InvalidProtocolInFinalMobileUrl,
    MalformedFinalMobileUrl,
    MissingHostInFinalMobileUrl,
    InvalidTldInFinalMobileUrl,
    InvalidFinalAppUrl,
    InvalidTagInFinalAppUrl,
    RedundantNestedFinalAppUrlTag,
    MultipleAppUrlsForOstype,
    InvalidOstype,
    InvalidProtocolForAppUrl,
    InvalidPackageIdForAppUrl,
    UrlCustomParametersCountExceedsLimit,
    InvalidCharactersInUrlCustomParameterKey,
    InvalidCharactersInUrlCustomParameterValue,
    InvalidTagInUrlCustomParameterValue,
    RedundantNestedUrlCustomParameterTag,
    MissingProtocol,
    InvalidProtocol,
    InvalidUrl,
    DestinationUrlDeprecated,
    InvalidTagInUrl,
    MissingUrlTag,
    DuplicateUrlId,
    InvalidUrlId,
    FinalUrlSuffixMalformed,
    InvalidTagInFinalUrlSuffix,
    InvalidTopLevelDomain,
    MalformedTopLevelDomain,
    MalformedUrl,
    MissingHost,
    NullCustomParameterValue,

}

Enum describing possible url field errors.

Variants

Unspecified

Enum unspecified.

Unknown

The received error code is not known in this version.

InvalidTrackingUrlTemplate

The tracking url template is invalid.

InvalidTagInTrackingUrlTemplate

The tracking url template contains invalid tag.

MissingTrackingUrlTemplateTag

The tracking url template must contain at least one tag (e.g. {lpurl}), This applies only to tracking url template associated with website ads or product ads.

MissingProtocolInTrackingUrlTemplate

The tracking url template must start with a valid protocol (or lpurl tag).

InvalidProtocolInTrackingUrlTemplate

The tracking url template starts with an invalid protocol.

MalformedTrackingUrlTemplate

The tracking url template contains illegal characters.

MissingHostInTrackingUrlTemplate

The tracking url template must contain a host name (or lpurl tag).

InvalidTldInTrackingUrlTemplate

The tracking url template has an invalid or missing top level domain extension.

RedundantNestedTrackingUrlTemplateTag

The tracking url template contains nested occurrences of the same conditional tag (i.e. {ifmobile:{ifmobile:x}}).

InvalidFinalUrl

The final url is invalid.

InvalidTagInFinalUrl

The final url contains invalid tag.

RedundantNestedFinalUrlTag

The final url contains nested occurrences of the same conditional tag (i.e. {ifmobile:{ifmobile:x}}).

MissingProtocolInFinalUrl

The final url must start with a valid protocol.

InvalidProtocolInFinalUrl

The final url starts with an invalid protocol.

MalformedFinalUrl

The final url contains illegal characters.

MissingHostInFinalUrl

The final url must contain a host name.

InvalidTldInFinalUrl

The tracking url template has an invalid or missing top level domain extension.

InvalidFinalMobileUrl

The final mobile url is invalid.

InvalidTagInFinalMobileUrl

The final mobile url contains invalid tag.

RedundantNestedFinalMobileUrlTag

The final mobile url contains nested occurrences of the same conditional tag (i.e. {ifmobile:{ifmobile:x}}).

MissingProtocolInFinalMobileUrl

The final mobile url must start with a valid protocol.

InvalidProtocolInFinalMobileUrl

The final mobile url starts with an invalid protocol.

MalformedFinalMobileUrl

The final mobile url contains illegal characters.

MissingHostInFinalMobileUrl

The final mobile url must contain a host name.

InvalidTldInFinalMobileUrl

The tracking url template has an invalid or missing top level domain extension.

InvalidFinalAppUrl

The final app url is invalid.

InvalidTagInFinalAppUrl

The final app url contains invalid tag.

RedundantNestedFinalAppUrlTag

The final app url contains nested occurrences of the same conditional tag (i.e. {ifmobile:{ifmobile:x}}).

MultipleAppUrlsForOstype

More than one app url found for the same OS type.

InvalidOstype

The OS type given for an app url is not valid.

InvalidProtocolForAppUrl

The protocol given for an app url is not valid. (E.g. “android-app://”)

InvalidPackageIdForAppUrl

The package id (app id) given for an app url is not valid.

UrlCustomParametersCountExceedsLimit

The number of url custom parameters for an resource exceeds the maximum limit allowed.

InvalidCharactersInUrlCustomParameterKey

An invalid character appears in the parameter key.

InvalidCharactersInUrlCustomParameterValue

An invalid character appears in the parameter value.

InvalidTagInUrlCustomParameterValue

The url custom parameter value fails url tag validation.

RedundantNestedUrlCustomParameterTag

The custom parameter contains nested occurrences of the same conditional tag (i.e. {ifmobile:{ifmobile:x}}).

MissingProtocol

The protocol (http:// or https://) is missing.

InvalidProtocol

Unsupported protocol in URL. Only http and https are supported.

InvalidUrl

The url is invalid.

DestinationUrlDeprecated

Destination Url is deprecated.

InvalidTagInUrl

The url contains invalid tag.

MissingUrlTag

The url must contain at least one tag (e.g. {lpurl}).

DuplicateUrlId

Duplicate url id.

InvalidUrlId

Invalid url id.

FinalUrlSuffixMalformed

The final url suffix cannot begin with ‘?’ or ‘&’ characters and must be a valid query string.

InvalidTagInFinalUrlSuffix

The final url suffix cannot contain {lpurl} related or {ignore} tags.

InvalidTopLevelDomain

The top level domain is invalid, e.g. not a public top level domain listed in publicsuffix.org.

MalformedTopLevelDomain

Malformed top level domain in URL.

MalformedUrl

Malformed URL.

MissingHost

No host found in URL.

NullCustomParameterValue

Custom parameter value cannot be null.

Implementations

impl UrlFieldError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of UrlFieldError.

pub fn from_i32(value: i32) -> Option<UrlFieldError>

Converts an i32 to a UrlFieldError, or None if value is not a valid variant.