Module google_api_proto::google::ads::searchads360::v0::common

Modules

value

Nested message and enum types in Value.

Structs

AdScheduleInfo

Represents an AdSchedule criterion.

AgeRangeInfo

An age range criterion.

AssetInteractionTarget

An AssetInteractionTarget segment.

AssetUsage

Contains the usage information of the asset.

AudienceInfo

An audience criterion.

BusinessProfileLocation

Business Profile location data synced from the linked Business Profile account.

CallToActionAsset

A call to action asset.

CustomParameter

A mapping that can be used by custom parameter tags in a tracking_url_template, final_urls, or mobile_final_urls.

DeviceInfo

A device criterion.

EnhancedCpc

An automated bidding strategy that raises bids for clicks that seem more likely to lead to a conversion and lowers them for clicks where they seem less likely.

FrequencyCapEntry

A rule specifying the maximum number of times an ad (or some set of ads) can be shown to a user over a particular time period.

GenderInfo

A gender criterion.

ImageAsset

An Image asset.

ImageDimension

Metadata for an image at a certain size, either original or resized.

Keyword

A Keyword criterion segment.

KeywordInfo

A keyword criterion.

LanguageInfo

A language criterion.

ListingGroupInfo

A listing group criterion.

LocationGroupInfo

A radius around a list of locations specified through a feed.

LocationInfo

A location criterion.

ManualCpa

Manual bidding strategy that allows advertiser to set the bid per advertiser-specified action.

ManualCpc

Manual click-based bidding where user pays per click.

ManualCpm

Manual impression-based bidding where user pays per thousand impressions.

MaximizeConversionValue

An automated bidding strategy to help get the most conversion value for your campaigns while spending your budget.

MaximizeConversions

An automated bidding strategy to help get the most conversions for your campaigns while spending your budget.

Metrics

Metrics data.

MobileAppAsset

An asset representing a mobile app.

PercentCpc

A bidding strategy where bids are a fraction of the advertised price for some good or service.

RealTimeBiddingSetting

Settings for Real-Time Bidding, a feature only available for campaigns targeting the Ad Exchange network.

SearchAds360ExpandedDynamicSearchAdInfo

An expanded dynamic search ad.

SearchAds360ExpandedTextAdInfo

A Search Ads 360 expanded text ad.

SearchAds360ProductAdInfo

A Search Ads 360 product ad.

SearchAds360ResponsiveSearchAdInfo

A Search Ads 360 responsive search ad.

SearchAds360TextAdInfo

A Search Ads 360 text ad.

Segments

Segment only fields.

TargetCpa

An automated bid strategy that sets bids to help get as many conversions as possible at the target cost-per-acquisition (CPA) you set.

TargetCpm

Target CPM (cost per thousand impressions) is an automated bidding strategy that sets bids to optimize performance given the target CPM you set.

TargetImpressionShare

An automated bidding strategy that sets bids so that a certain percentage of search ads are shown at the top of the first page (or other targeted location).

TargetOutrankShare

An automated bidding strategy that sets bids based on the target fraction of auctions where the advertiser should outrank a specific competitor. This strategy is deprecated.

TargetRestriction

The list of per-targeting-dimension targeting settings.

TargetRoas

An automated bidding strategy that helps you maximize revenue while averaging a specific target return on ad spend (ROAS).

TargetSpend

An automated bid strategy that sets your bids to help get as many clicks as possible within your budget.

TargetingSetting

Settings for the targeting-related features, at the campaign and ad group levels. For more details about the targeting setting, visit https://support.google.com/google-ads/answer/7365594

TextAsset

A Text asset.

TextLabel

A type of label displaying text on a colored background.

UnifiedCallAsset

A unified call asset.

UnifiedCalloutAsset

A unified callout asset.

UnifiedLocationAsset

A unified location asset.

UnifiedPageFeedAsset

A Unified Page Feed asset.

UnifiedSitelinkAsset

A unified sitelink asset.

UserListInfo

A User List criterion. Represents a user list that is defined by the advertiser to be targeted.

Value

A generic data container.

WebpageConditionInfo

Logical expression for targeting webpages of an advertiser’s website.

WebpageInfo

Represents a criterion for targeting webpages of an advertiser’s website.

YoutubeVideoAsset

A YouTube asset.