Struct google_api_proto::google::ads::searchads360::v0::common::Metrics

pub struct Metrics {

    pub absolute_top_impression_percentage: Option<f64>,
    pub all_conversions_from_interactions_rate: Option<f64>,
    pub all_conversions_value: Option<f64>,
    pub all_conversions_value_by_conversion_date: f64,
    pub all_conversions: Option<f64>,
    pub all_conversions_by_conversion_date: f64,
    pub all_conversions_value_per_cost: Option<f64>,
    pub all_conversions_from_click_to_call: Option<f64>,
    pub all_conversions_from_directions: Option<f64>,
    pub all_conversions_from_interactions_value_per_interaction: Option<f64>,
    pub all_conversions_from_menu: Option<f64>,
    pub all_conversions_from_order: Option<f64>,
    pub all_conversions_from_other_engagement: Option<f64>,
    pub all_conversions_from_store_visit: Option<f64>,
    pub visits: Option<f64>,
    pub all_conversions_from_store_website: Option<f64>,
    pub average_cost: Option<f64>,
    pub average_cpc: Option<f64>,
    pub average_cpm: Option<f64>,
    pub clicks: Option<i64>,
    pub content_budget_lost_impression_share: Option<f64>,
    pub content_impression_share: Option<f64>,
    pub conversion_custom_metrics: Vec<Value>,
    pub content_rank_lost_impression_share: Option<f64>,
    pub conversions_from_interactions_rate: Option<f64>,
    pub client_account_conversions_value: Option<f64>,
    pub conversions_value_by_conversion_date: f64,
    pub conversions_value_per_cost: Option<f64>,
    pub conversions_from_interactions_value_per_interaction: Option<f64>,
    pub client_account_conversions: Option<f64>,
    pub conversions_by_conversion_date: f64,
    pub cost_micros: Option<i64>,
    pub cost_per_all_conversions: Option<f64>,
    pub cost_per_conversion: Option<f64>,
    pub cost_per_current_model_attributed_conversion: Option<f64>,
    pub cross_device_conversions: Option<f64>,
    pub cross_device_conversions_value: Option<f64>,
    pub ctr: Option<f64>,
    pub conversions: Option<f64>,
    pub conversions_value: Option<f64>,
    pub historical_creative_quality_score: i32,
    pub average_quality_score: Option<f64>,
    pub historical_landing_page_quality_score: i32,
    pub historical_quality_score: Option<i64>,
    pub historical_search_predicted_ctr: i32,
    pub impressions: Option<i64>,
    pub interaction_rate: Option<f64>,
    pub interactions: Option<i64>,
    pub interaction_event_types: Vec<i32>,
    pub invalid_click_rate: Option<f64>,
    pub invalid_clicks: Option<i64>,
    pub mobile_friendly_clicks_percentage: Option<f64>,
    pub raw_event_conversion_metrics: Vec<Value>,
    pub search_absolute_top_impression_share: Option<f64>,
    pub search_budget_lost_absolute_top_impression_share: Option<f64>,
    pub search_budget_lost_impression_share: Option<f64>,
    pub search_budget_lost_top_impression_share: Option<f64>,
    pub search_click_share: Option<f64>,
    pub search_exact_match_impression_share: Option<f64>,
    pub search_impression_share: Option<f64>,
    pub search_rank_lost_absolute_top_impression_share: Option<f64>,
    pub search_rank_lost_impression_share: Option<f64>,
    pub search_rank_lost_top_impression_share: Option<f64>,
    pub search_top_impression_share: Option<f64>,
    pub top_impression_percentage: Option<f64>,
    pub value_per_all_conversions: Option<f64>,
    pub value_per_all_conversions_by_conversion_date: Option<f64>,
    pub value_per_conversion: Option<f64>,
    pub value_per_conversions_by_conversion_date: Option<f64>,
    pub client_account_view_through_conversions: Option<i64>,
    pub client_account_cross_sell_cost_of_goods_sold_micros: Option<i64>,
    pub cross_sell_cost_of_goods_sold_micros: Option<i64>,
    pub client_account_cross_sell_gross_profit_micros: Option<i64>,
    pub cross_sell_gross_profit_micros: Option<i64>,
    pub client_account_cross_sell_revenue_micros: Option<i64>,
    pub cross_sell_revenue_micros: Option<i64>,
    pub client_account_cross_sell_units_sold: Option<f64>,
    pub cross_sell_units_sold: Option<f64>,
    pub client_account_lead_cost_of_goods_sold_micros: Option<i64>,
    pub lead_cost_of_goods_sold_micros: Option<i64>,
    pub client_account_lead_gross_profit_micros: Option<i64>,
    pub lead_gross_profit_micros: Option<i64>,
    pub client_account_lead_revenue_micros: Option<i64>,
    pub lead_revenue_micros: Option<i64>,
    pub client_account_lead_units_sold: Option<f64>,
    pub lead_units_sold: Option<f64>,

}

Metrics data.

Fields
absolute_top_impression_percentage: Option<f64>

Search absolute top impression share is the percentage of your Search ad impressions that are shown in the most prominent Search position.

all_conversions_from_interactions_rate: Option<f64>

All conversions from interactions (as oppose to view through conversions) divided by the number of ad interactions.

all_conversions_value: Option<f64>

The value of all conversions.

all_conversions_value_by_conversion_date: f64

The value of all conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/sa360/answer/9250611.

all_conversions: Option<f64>

The total number of conversions. This includes all conversions regardless of the value of include_in_conversions_metric.

all_conversions_by_conversion_date: f64

The total number of conversions. This includes all conversions regardless of the value of include_in_conversions_metric. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/sa360/answer/9250611.

all_conversions_value_per_cost: Option<f64>

The value of all conversions divided by the total cost of ad interactions (such as clicks for text ads or views for video ads).

all_conversions_from_click_to_call: Option<f64>

The number of times people clicked the “Call” button to call a store during or after clicking an ad. This number doesn’t include whether or not calls were connected, or the duration of any calls.

This metric applies to feed items only.

all_conversions_from_directions: Option<f64>

The number of times people clicked a “Get directions” button to navigate to a store after clicking an ad.

This metric applies to feed items only.

all_conversions_from_interactions_value_per_interaction: Option<f64>

The value of all conversions from interactions divided by the total number of interactions.

all_conversions_from_menu: Option<f64>

The number of times people clicked a link to view a store’s menu after clicking an ad.

This metric applies to feed items only.

all_conversions_from_order: Option<f64>

The number of times people placed an order at a store after clicking an ad.

This metric applies to feed items only.

all_conversions_from_other_engagement: Option<f64>

The number of other conversions (for example, posting a review or saving a location for a store) that occurred after people clicked an ad.

This metric applies to feed items only.

all_conversions_from_store_visit: Option<f64>

Estimated number of times people visited a store after clicking an ad.

This metric applies to feed items only.

visits: Option<f64>

Clicks that Search Ads 360 has successfully recorded and forwarded to an advertiser’s landing page.

all_conversions_from_store_website: Option<f64>

The number of times that people were taken to a store’s URL after clicking an ad.

This metric applies to feed items only.

average_cost: Option<f64>

The average amount you pay per interaction. This amount is the total cost of your ads divided by the total number of interactions.

average_cpc: Option<f64>

The total cost of all clicks divided by the total number of clicks received. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

average_cpm: Option<f64>

Average cost-per-thousand impressions (CPM). This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

clicks: Option<i64>

The number of clicks.

content_budget_lost_impression_share: Option<f64>

The estimated percent of times that your ad was eligible to show on the Display Network but didn’t because your budget was too low. Note: Content budget lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

content_impression_share: Option<f64>

The impressions you’ve received on the Display Network divided by the estimated number of impressions you were eligible to receive. Note: Content impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

conversion_custom_metrics: Vec<Value>

The conversion custom metrics.

content_rank_lost_impression_share: Option<f64>

The estimated percentage of impressions on the Display Network that your ads didn’t receive due to poor Ad Rank. Note: Content rank lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

conversions_from_interactions_rate: Option<f64>

Average biddable conversions (from interaction) per conversion eligible interaction. Shows how often, on average, an ad interaction leads to a biddable conversion.

client_account_conversions_value: Option<f64>

The value of client account conversions. This only includes conversion actions which include_in_client_account_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_value_by_conversion_date: f64

The sum of biddable conversions value by conversion date. When this column is selected with date, the values in date column means the conversion date.

conversions_value_per_cost: Option<f64>

The value of biddable conversion divided by the total cost of conversion eligible interactions.

conversions_from_interactions_value_per_interaction: Option<f64>

The value of conversions from interactions divided by the number of ad interactions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

client_account_conversions: Option<f64>

The number of client account conversions. This only includes conversion actions which include_in_client_account_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_by_conversion_date: f64

The sum of conversions by conversion date for biddable conversion types. Can be fractional due to attribution modeling. When this column is selected with date, the values in date column means the conversion date.

cost_micros: Option<i64>

The sum of your cost-per-click (CPC) and cost-per-thousand impressions (CPM) costs during this period. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

cost_per_all_conversions: Option<f64>

The cost of ad interactions divided by all conversions.

cost_per_conversion: Option<f64>

Average conversion eligible cost per biddable conversion.

cost_per_current_model_attributed_conversion: Option<f64>

The cost of ad interactions divided by current model attributed conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

cross_device_conversions: Option<f64>

Conversions from when a customer clicks on an ad on one device, then converts on a different device or browser. Cross-device conversions are already included in all_conversions.

cross_device_conversions_value: Option<f64>

The sum of the value of cross-device conversions.

ctr: Option<f64>

The number of clicks your ad receives (Clicks) divided by the number of times your ad is shown (Impressions).

conversions: Option<f64>

The number of conversions. This only includes conversion actions which include_in_conversions_metric attribute is set to true. If you use conversion-based bidding, your bid strategies will optimize for these conversions.

conversions_value: Option<f64>

The sum of conversion values for the conversions included in the “conversions” field. This metric is useful only if you entered a value for your conversion actions.

historical_creative_quality_score: i32

The creative historical quality score.

average_quality_score: Option<f64>

The average quality score.

historical_landing_page_quality_score: i32

The quality of historical landing page experience.

historical_quality_score: Option<i64>

The historical quality score.

historical_search_predicted_ctr: i32

The historical search predicted click through rate (CTR).

impressions: Option<i64>

Count of how often your ad has appeared on a search results page or website on the Google Network.

interaction_rate: Option<f64>

How often people interact with your ad after it is shown to them. This is the number of interactions divided by the number of times your ad is shown.

interactions: Option<i64>

The number of interactions. An interaction is the main user action associated with an ad format-clicks for text and shopping ads, views for video ads, and so on.

interaction_event_types: Vec<i32>

The types of payable and free interactions.

invalid_click_rate: Option<f64>

The percentage of clicks filtered out of your total number of clicks (filtered + non-filtered clicks) during the reporting period.

invalid_clicks: Option<i64>

Number of clicks Google considers illegitimate and doesn’t charge you for.

mobile_friendly_clicks_percentage: Option<f64>

The percentage of mobile clicks that go to a mobile-friendly page.

raw_event_conversion_metrics: Vec<Value>

The raw event conversion metrics.

search_absolute_top_impression_share: Option<f64>

The percentage of the customer’s Shopping or Search ad impressions that are shown in the most prominent Shopping position. See https://support.google.com/sa360/answer/9566729 for details. Any value below 0.1 is reported as 0.0999.

search_budget_lost_absolute_top_impression_share: Option<f64>

The number estimating how often your ad wasn’t the very first ad among the top ads in the search results due to a low budget. Note: Search budget lost absolute top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_budget_lost_impression_share: Option<f64>

The estimated percent of times that your ad was eligible to show on the Search Network but didn’t because your budget was too low. Note: Search budget lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_budget_lost_top_impression_share: Option<f64>

The number estimating how often your ad didn’t show adjacent to the top organic search results due to a low budget. Note: Search budget lost top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_click_share: Option<f64>

The number of clicks you’ve received on the Search Network divided by the estimated number of clicks you were eligible to receive. Note: Search click share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_exact_match_impression_share: Option<f64>

The impressions you’ve received divided by the estimated number of impressions you were eligible to receive on the Search Network for search terms that matched your keywords exactly (or were close variants of your keyword), regardless of your keyword match types. Note: Search exact match impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_impression_share: Option<f64>

The impressions you’ve received on the Search Network divided by the estimated number of impressions you were eligible to receive. Note: Search impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

search_rank_lost_absolute_top_impression_share: Option<f64>

The number estimating how often your ad wasn’t the very first ad among the top ads in the search results due to poor Ad Rank. Note: Search rank lost absolute top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_rank_lost_impression_share: Option<f64>

The estimated percentage of impressions on the Search Network that your ads didn’t receive due to poor Ad Rank. Note: Search rank lost impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_rank_lost_top_impression_share: Option<f64>

The number estimating how often your ad didn’t show adjacent to the top organic search results due to poor Ad Rank. Note: Search rank lost top impression share is reported in the range of 0 to 0.9. Any value above 0.9 is reported as 0.9001.

search_top_impression_share: Option<f64>

The impressions you’ve received among the top ads compared to the estimated number of impressions you were eligible to receive among the top ads. Note: Search top impression share is reported in the range of 0.1 to 1. Any value below 0.1 is reported as 0.0999.

Top ads are generally above the top organic results, although they may show below the top organic results on certain queries.

top_impression_percentage: Option<f64>

The percent of your ad impressions that are shown adjacent to the top organic search results.

value_per_all_conversions: Option<f64>

The value of all conversions divided by the number of all conversions.

value_per_all_conversions_by_conversion_date: Option<f64>

The value of all conversions divided by the number of all conversions. When this column is selected with date, the values in date column means the conversion date. Details for the by_conversion_date columns are available at https://support.google.com/sa360/answer/9250611.

value_per_conversion: Option<f64>

The value of biddable conversion divided by the number of biddable conversions. Shows how much, on average, each of the biddable conversions is worth.

value_per_conversions_by_conversion_date: Option<f64>

Biddable conversions value by conversion date divided by biddable conversions by conversion date. Shows how much, on average, each of the biddable conversions is worth (by conversion date). When this column is selected with date, the values in date column means the conversion date.

client_account_view_through_conversions: Option<i64>

The total number of view-through conversions. These happen when a customer sees an image or rich media ad, then later completes a conversion on your site without interacting with (for example, clicking on) another ad.

client_account_cross_sell_cost_of_goods_sold_micros: Option<i64>

Client account cross-sell cost of goods sold (COGS) is the total cost of products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell cost of goods sold is the total cost of the products sold that weren’t advertised. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat has a cost of goods sold value of $3, the shirt has a cost of goods sold value of $5. The cross-sell cost of goods sold for this order is $5. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

cross_sell_cost_of_goods_sold_micros: Option<i64>

Cross-sell cost of goods sold (COGS) is the total cost of products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell cost of goods sold is the total cost of the products sold that weren’t advertised. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat has a cost of goods sold value of $3, the shirt has a cost of goods sold value of $5. The cross-sell cost of goods sold for this order is $5. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_cross_sell_gross_profit_micros: Option<i64>

Client account cross-sell gross profit is the profit you made from products sold as a result of advertising a different product, minus cost of goods sold (COGS). How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the purchase is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell gross profit is the revenue you made from cross-sell attributed to your ads minus the cost of the goods sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The shirt is priced $20 and has a cost of goods sold value of $5. The cross-sell gross profit of this order is $15 = $20 - $5. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

cross_sell_gross_profit_micros: Option<i64>

Cross-sell gross profit is the profit you made from products sold as a result of advertising a different product, minus cost of goods sold (COGS). How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the purchase is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell gross profit is the revenue you made from cross-sell attributed to your ads minus the cost of the goods sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The shirt is priced $20 and has a cost of goods sold value of $5. The cross-sell gross profit of this order is $15 = $20 - $5. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_cross_sell_revenue_micros: Option<i64>

Client account cross-sell revenue is the total amount you made from products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell revenue is the total value you made from cross-sell attributed to your ads. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and the shirt is priced $20. The cross-sell revenue of this order is $20. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

cross_sell_revenue_micros: Option<i64>

Cross-sell revenue is the total amount you made from products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell revenue is the total value you made from cross-sell attributed to your ads. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and the shirt is priced $20. The cross-sell revenue of this order is $20. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_cross_sell_units_sold: Option<f64>

Client account cross-sell units sold is the total number of products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell units sold is the total number of cross-sold products from all orders attributed to your ads. Example: Someone clicked on a Shopping ad for a hat then bought the same hat, a shirt and a jacket. The cross-sell units sold in this order is 2. This metric is only available if you report conversions with cart data.

cross_sell_units_sold: Option<f64>

Cross-sell units sold is the total number of products sold as a result of advertising a different product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If these products don’t match then this is considered cross-sell. Cross-sell units sold is the total number of cross-sold products from all orders attributed to your ads. Example: Someone clicked on a Shopping ad for a hat then bought the same hat, a shirt and a jacket. The cross-sell units sold in this order is 2. This metric is only available if you report conversions with cart data.

client_account_lead_cost_of_goods_sold_micros: Option<i64>

Client account lead cost of goods sold (COGS) is the total cost of products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the cost of these goods is counted under lead cost of goods sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat has a cost of goods sold value of $3, the shirt has a cost of goods sold value of $5. The lead cost of goods sold for this order is $3. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

lead_cost_of_goods_sold_micros: Option<i64>

Lead cost of goods sold (COGS) is the total cost of products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the cost of these goods is counted under lead cost of goods sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat has a cost of goods sold value of $3, the shirt has a cost of goods sold value of $5. The lead cost of goods sold for this order is $3. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_lead_gross_profit_micros: Option<i64>

Client account lead gross profit is the profit you made from products sold as a result of advertising the same product, minus cost of goods sold (COGS). How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the revenue you made from these sales minus the cost of goods sold is your lead gross profit. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and has a cost of goods sold value of $3. The lead gross profit of this order is $7 = $10 - $3. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

lead_gross_profit_micros: Option<i64>

Lead gross profit is the profit you made from products sold as a result of advertising the same product, minus cost of goods sold (COGS). How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the revenue you made from these sales minus the cost of goods sold is your lead gross profit. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and has a cost of goods sold value of $3. The lead gross profit of this order is $7 = $10 - $3. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_lead_revenue_micros: Option<i64>

Client account lead revenue is the total amount you made from products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the total value you made from the sales of these products is shown under lead revenue. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and the shirt is priced $20. The lead revenue of this order is $10. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

lead_revenue_micros: Option<i64>

Lead revenue is the total amount you made from products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the total value you made from the sales of these products is shown under lead revenue. Example: Someone clicked on a Shopping ad for a hat then bought the same hat and a shirt. The hat is priced $10 and the shirt is priced $20. The lead revenue of this order is $10. This metric is only available if you report conversions with cart data. This metric is a monetary value and returned in the customer’s currency by default. See the metrics_currency parameter at https://developers.google.com/search-ads/reporting/query/query-structure#parameters_clause

client_account_lead_units_sold: Option<f64>

Client account lead units sold is the total number of products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the total number of these products sold is shown under lead units sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat, a shirt and a jacket. The lead units sold in this order is 1. This metric is only available if you report conversions with cart data.

lead_units_sold: Option<f64>

Lead units sold is the total number of products sold as a result of advertising the same product. How it works: You report conversions with cart data for completed purchases on your website. If the ad that was interacted with before the purchase has an associated product (see Shopping Ads) then this product is considered the advertised product. Any product included in the order the customer places is a sold product. If the advertised and sold products match, then the total number of these products sold is shown under lead units sold. Example: Someone clicked on a Shopping ad for a hat then bought the same hat, a shirt and a jacket. The lead units sold in this order is 1. This metric is only available if you report conversions with cart data.

Implementations

impl Metrics

pub fn historical_creative_quality_score(&self) -> QualityScoreBucket

Returns the enum value of historical_creative_quality_score, or the default if the field is set to an invalid enum value.

pub fn set_historical_creative_quality_score( &mut self, value: QualityScoreBucket, )

Sets historical_creative_quality_score to the provided enum value.

pub fn historical_landing_page_quality_score(&self) -> QualityScoreBucket

Returns the enum value of historical_landing_page_quality_score, or the default if the field is set to an invalid enum value.

pub fn set_historical_landing_page_quality_score( &mut self, value: QualityScoreBucket, )

Sets historical_landing_page_quality_score to the provided enum value.

pub fn historical_search_predicted_ctr(&self) -> QualityScoreBucket

Returns the enum value of historical_search_predicted_ctr, or the default if the field is set to an invalid enum value.

pub fn set_historical_search_predicted_ctr(&mut self, value: QualityScoreBucket)

Sets historical_search_predicted_ctr to the provided enum value.

pub fn interaction_event_types( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<InteractionEventType>>

Returns an iterator which yields the valid enum values contained in interaction_event_types.

pub fn push_interaction_event_types(&mut self, value: InteractionEventType)

Appends the provided enum value to interaction_event_types.

pub fn clicks(&self) -> i64

Returns the value of clicks, or the default value if clicks is unset.

pub fn search_absolute_top_impression_share(&self) -> f64

Returns the value of search_absolute_top_impression_share, or the default value if search_absolute_top_impression_share is unset.

pub fn search_budget_lost_absolute_top_impression_share(&self) -> f64

Returns the value of search_budget_lost_absolute_top_impression_share, or the default value if search_budget_lost_absolute_top_impression_share is unset.

pub fn search_budget_lost_impression_share(&self) -> f64

Returns the value of search_budget_lost_impression_share, or the default value if search_budget_lost_impression_share is unset.

pub fn search_budget_lost_top_impression_share(&self) -> f64

Returns the value of search_budget_lost_top_impression_share, or the default value if search_budget_lost_top_impression_share is unset.

pub fn search_click_share(&self) -> f64

Returns the value of search_click_share, or the default value if search_click_share is unset.

pub fn search_exact_match_impression_share(&self) -> f64

Returns the value of search_exact_match_impression_share, or the default value if search_exact_match_impression_share is unset.

pub fn search_impression_share(&self) -> f64

Returns the value of search_impression_share, or the default value if search_impression_share is unset.

pub fn search_rank_lost_absolute_top_impression_share(&self) -> f64

Returns the value of search_rank_lost_absolute_top_impression_share, or the default value if search_rank_lost_absolute_top_impression_share is unset.

pub fn search_rank_lost_impression_share(&self) -> f64

Returns the value of search_rank_lost_impression_share, or the default value if search_rank_lost_impression_share is unset.

pub fn search_rank_lost_top_impression_share(&self) -> f64

Returns the value of search_rank_lost_top_impression_share, or the default value if search_rank_lost_top_impression_share is unset.

pub fn search_top_impression_share(&self) -> f64

Returns the value of search_top_impression_share, or the default value if search_top_impression_share is unset.

pub fn top_impression_percentage(&self) -> f64

Returns the value of top_impression_percentage, or the default value if top_impression_percentage is unset.

pub fn value_per_all_conversions(&self) -> f64

Returns the value of value_per_all_conversions, or the default value if value_per_all_conversions is unset.

pub fn client_account_view_through_conversions(&self) -> i64

Returns the value of client_account_view_through_conversions, or the default value if client_account_view_through_conversions is unset.

pub fn content_budget_lost_impression_share(&self) -> f64

Returns the value of content_budget_lost_impression_share, or the default value if content_budget_lost_impression_share is unset.

pub fn content_impression_share(&self) -> f64

Returns the value of content_impression_share, or the default value if content_impression_share is unset.

pub fn content_rank_lost_impression_share(&self) -> f64

Returns the value of content_rank_lost_impression_share, or the default value if content_rank_lost_impression_share is unset.

pub fn client_account_conversions_value(&self) -> f64

Returns the value of client_account_conversions_value, or the default value if client_account_conversions_value is unset.

pub fn conversions_from_interactions_value_per_interaction(&self) -> f64

Returns the value of conversions_from_interactions_value_per_interaction, or the default value if conversions_from_interactions_value_per_interaction is unset.

pub fn client_account_conversions(&self) -> f64

Returns the value of client_account_conversions, or the default value if client_account_conversions is unset.

pub fn cost_per_all_conversions(&self) -> f64

Returns the value of cost_per_all_conversions, or the default value if cost_per_all_conversions is unset.

pub fn cost_per_current_model_attributed_conversion(&self) -> f64

Returns the value of cost_per_current_model_attributed_conversion, or the default value if cost_per_current_model_attributed_conversion is unset.

pub fn cross_device_conversions(&self) -> f64

Returns the value of cross_device_conversions, or the default value if cross_device_conversions is unset.

pub fn ctr(&self) -> f64

Returns the value of ctr, or the default value if ctr is unset.

pub fn absolute_top_impression_percentage(&self) -> f64

Returns the value of absolute_top_impression_percentage, or the default value if absolute_top_impression_percentage is unset.

pub fn all_conversions_from_interactions_rate(&self) -> f64

Returns the value of all_conversions_from_interactions_rate, or the default value if all_conversions_from_interactions_rate is unset.

pub fn all_conversions_value(&self) -> f64

Returns the value of all_conversions_value, or the default value if all_conversions_value is unset.

pub fn all_conversions(&self) -> f64

Returns the value of all_conversions, or the default value if all_conversions is unset.

pub fn all_conversions_value_per_cost(&self) -> f64

Returns the value of all_conversions_value_per_cost, or the default value if all_conversions_value_per_cost is unset.

pub fn all_conversions_from_click_to_call(&self) -> f64

Returns the value of all_conversions_from_click_to_call, or the default value if all_conversions_from_click_to_call is unset.

pub fn all_conversions_from_directions(&self) -> f64

Returns the value of all_conversions_from_directions, or the default value if all_conversions_from_directions is unset.

pub fn all_conversions_from_interactions_value_per_interaction(&self) -> f64

Returns the value of all_conversions_from_interactions_value_per_interaction, or the default value if all_conversions_from_interactions_value_per_interaction is unset.

pub fn all_conversions_from_menu(&self) -> f64

Returns the value of all_conversions_from_menu, or the default value if all_conversions_from_menu is unset.

pub fn all_conversions_from_order(&self) -> f64

Returns the value of all_conversions_from_order, or the default value if all_conversions_from_order is unset.

pub fn all_conversions_from_other_engagement(&self) -> f64

Returns the value of all_conversions_from_other_engagement, or the default value if all_conversions_from_other_engagement is unset.

pub fn all_conversions_from_store_visit(&self) -> f64

Returns the value of all_conversions_from_store_visit, or the default value if all_conversions_from_store_visit is unset.

pub fn all_conversions_from_store_website(&self) -> f64

Returns the value of all_conversions_from_store_website, or the default value if all_conversions_from_store_website is unset.

pub fn average_cost(&self) -> f64

Returns the value of average_cost, or the default value if average_cost is unset.

pub fn historical_quality_score(&self) -> i64

Returns the value of historical_quality_score, or the default value if historical_quality_score is unset.

pub fn impressions(&self) -> i64

Returns the value of impressions, or the default value if impressions is unset.

pub fn interaction_rate(&self) -> f64

Returns the value of interaction_rate, or the default value if interaction_rate is unset.

pub fn interactions(&self) -> i64

Returns the value of interactions, or the default value if interactions is unset.

pub fn invalid_click_rate(&self) -> f64

Returns the value of invalid_click_rate, or the default value if invalid_click_rate is unset.

pub fn invalid_clicks(&self) -> i64

Returns the value of invalid_clicks, or the default value if invalid_clicks is unset.

pub fn mobile_friendly_clicks_percentage(&self) -> f64

Returns the value of mobile_friendly_clicks_percentage, or the default value if mobile_friendly_clicks_percentage is unset.

pub fn value_per_all_conversions_by_conversion_date(&self) -> f64

Returns the value of value_per_all_conversions_by_conversion_date, or the default value if value_per_all_conversions_by_conversion_date is unset.

pub fn conversions(&self) -> f64

Returns the value of conversions, or the default value if conversions is unset.

pub fn conversions_value(&self) -> f64

Returns the value of conversions_value, or the default value if conversions_value is unset.

pub fn cross_device_conversions_value(&self) -> f64

Returns the value of cross_device_conversions_value, or the default value if cross_device_conversions_value is unset.

pub fn conversions_from_interactions_rate(&self) -> f64

Returns the value of conversions_from_interactions_rate, or the default value if conversions_from_interactions_rate is unset.

pub fn value_per_conversions_by_conversion_date(&self) -> f64

Returns the value of value_per_conversions_by_conversion_date, or the default value if value_per_conversions_by_conversion_date is unset.

pub fn cost_per_conversion(&self) -> f64

Returns the value of cost_per_conversion, or the default value if cost_per_conversion is unset.

pub fn value_per_conversion(&self) -> f64

Returns the value of value_per_conversion, or the default value if value_per_conversion is unset.

pub fn conversions_value_per_cost(&self) -> f64

Returns the value of conversions_value_per_cost, or the default value if conversions_value_per_cost is unset.

pub fn visits(&self) -> f64

Returns the value of visits, or the default value if visits is unset.

pub fn client_account_cross_sell_units_sold(&self) -> f64

Returns the value of client_account_cross_sell_units_sold, or the default value if client_account_cross_sell_units_sold is unset.

pub fn client_account_lead_units_sold(&self) -> f64

Returns the value of client_account_lead_units_sold, or the default value if client_account_lead_units_sold is unset.

pub fn cost_micros(&self) -> i64

Returns the value of cost_micros, or the default value if cost_micros is unset.

pub fn average_cpc(&self) -> f64

Returns the value of average_cpc, or the default value if average_cpc is unset.

pub fn average_cpm(&self) -> f64

Returns the value of average_cpm, or the default value if average_cpm is unset.

pub fn client_account_cross_sell_cost_of_goods_sold_micros(&self) -> i64

Returns the value of client_account_cross_sell_cost_of_goods_sold_micros, or the default value if client_account_cross_sell_cost_of_goods_sold_micros is unset.

pub fn client_account_cross_sell_gross_profit_micros(&self) -> i64

Returns the value of client_account_cross_sell_gross_profit_micros, or the default value if client_account_cross_sell_gross_profit_micros is unset.

pub fn client_account_cross_sell_revenue_micros(&self) -> i64

Returns the value of client_account_cross_sell_revenue_micros, or the default value if client_account_cross_sell_revenue_micros is unset.

pub fn client_account_lead_cost_of_goods_sold_micros(&self) -> i64

Returns the value of client_account_lead_cost_of_goods_sold_micros, or the default value if client_account_lead_cost_of_goods_sold_micros is unset.

pub fn client_account_lead_gross_profit_micros(&self) -> i64

Returns the value of client_account_lead_gross_profit_micros, or the default value if client_account_lead_gross_profit_micros is unset.

pub fn client_account_lead_revenue_micros(&self) -> i64

Returns the value of client_account_lead_revenue_micros, or the default value if client_account_lead_revenue_micros is unset.

pub fn cross_sell_cost_of_goods_sold_micros(&self) -> i64

Returns the value of cross_sell_cost_of_goods_sold_micros, or the default value if cross_sell_cost_of_goods_sold_micros is unset.

pub fn cross_sell_gross_profit_micros(&self) -> i64

Returns the value of cross_sell_gross_profit_micros, or the default value if cross_sell_gross_profit_micros is unset.

pub fn cross_sell_revenue_micros(&self) -> i64

Returns the value of cross_sell_revenue_micros, or the default value if cross_sell_revenue_micros is unset.

pub fn cross_sell_units_sold(&self) -> f64

Returns the value of cross_sell_units_sold, or the default value if cross_sell_units_sold is unset.

pub fn lead_cost_of_goods_sold_micros(&self) -> i64

Returns the value of lead_cost_of_goods_sold_micros, or the default value if lead_cost_of_goods_sold_micros is unset.

pub fn lead_gross_profit_micros(&self) -> i64

Returns the value of lead_gross_profit_micros, or the default value if lead_gross_profit_micros is unset.

pub fn lead_revenue_micros(&self) -> i64

Returns the value of lead_revenue_micros, or the default value if lead_revenue_micros is unset.

pub fn lead_units_sold(&self) -> f64

Returns the value of lead_units_sold, or the default value if lead_units_sold is unset.

pub fn average_quality_score(&self) -> f64

Returns the value of average_quality_score, or the default value if average_quality_score is unset.