Module google_api_proto::google::ads::searchads360::v0::enums

Modules

account_status_enum

Nested message and enum types in AccountStatusEnum.

account_type_enum

Nested message and enum types in AccountTypeEnum.

ad_group_ad_engine_status_enum

Nested message and enum types in AdGroupAdEngineStatusEnum.

ad_group_ad_rotation_mode_enum

Nested message and enum types in AdGroupAdRotationModeEnum.

ad_group_ad_status_enum

Nested message and enum types in AdGroupAdStatusEnum.

ad_group_criterion_engine_status_enum

Nested message and enum types in AdGroupCriterionEngineStatusEnum.

ad_group_criterion_status_enum

Nested message and enum types in AdGroupCriterionStatusEnum.

ad_group_engine_status_enum

Nested message and enum types in AdGroupEngineStatusEnum.

ad_group_status_enum

Nested message and enum types in AdGroupStatusEnum.

ad_group_type_enum

Nested message and enum types in AdGroupTypeEnum.

ad_network_type_enum

Nested message and enum types in AdNetworkTypeEnum.

ad_serving_optimization_status_enum

Nested message and enum types in AdServingOptimizationStatusEnum.

ad_strength_enum

Nested message and enum types in AdStrengthEnum.

ad_type_enum

Nested message and enum types in AdTypeEnum.

advertising_channel_sub_type_enum

Nested message and enum types in AdvertisingChannelSubTypeEnum.

advertising_channel_type_enum

Nested message and enum types in AdvertisingChannelTypeEnum.

age_range_type_enum

Nested message and enum types in AgeRangeTypeEnum.

asset_engine_status_enum

Nested message and enum types in AssetEngineStatusEnum.

asset_field_type_enum

Nested message and enum types in AssetFieldTypeEnum.

asset_group_status_enum

Nested message and enum types in AssetGroupStatusEnum.

asset_link_status_enum

Nested message and enum types in AssetLinkStatusEnum.

asset_set_asset_status_enum

Nested message and enum types in AssetSetAssetStatusEnum.

asset_set_link_status_enum

Nested message and enum types in AssetSetLinkStatusEnum.

asset_status_enum

Nested message and enum types in AssetStatusEnum.

asset_type_enum

Nested message and enum types in AssetTypeEnum.

attribution_model_enum

Nested message and enum types in AttributionModelEnum.

attribution_type_enum

Nested message and enum types in AttributionTypeEnum.

bidding_strategy_status_enum

Nested message and enum types in BiddingStrategyStatusEnum.

bidding_strategy_system_status_enum

Nested message and enum types in BiddingStrategySystemStatusEnum.

bidding_strategy_type_enum

Nested message and enum types in BiddingStrategyTypeEnum.

budget_delivery_method_enum

Nested message and enum types in BudgetDeliveryMethodEnum.

budget_period_enum

Nested message and enum types in BudgetPeriodEnum.

call_conversion_reporting_state_enum

Nested message and enum types in CallConversionReportingStateEnum.

call_to_action_type_enum

Nested message and enum types in CallToActionTypeEnum.

campaign_criterion_status_enum

Nested message and enum types in CampaignCriterionStatusEnum.

campaign_serving_status_enum

Nested message and enum types in CampaignServingStatusEnum.

campaign_status_enum

Nested message and enum types in CampaignStatusEnum.

conversion_action_category_enum

Nested message and enum types in ConversionActionCategoryEnum.

conversion_action_status_enum

Nested message and enum types in ConversionActionStatusEnum.

conversion_action_type_enum

Nested message and enum types in ConversionActionTypeEnum.

conversion_custom_variable_cardinality_enum

Nested message and enum types in ConversionCustomVariableCardinalityEnum.

conversion_custom_variable_family_enum

Nested message and enum types in ConversionCustomVariableFamilyEnum.

conversion_custom_variable_status_enum

Nested message and enum types in ConversionCustomVariableStatusEnum.

conversion_status_enum

Nested message and enum types in ConversionStatusEnum.

conversion_tracking_status_enum

Nested message and enum types in ConversionTrackingStatusEnum.

criterion_type_enum

Nested message and enum types in CriterionTypeEnum.

custom_column_render_type_enum

Nested message and enum types in CustomColumnRenderTypeEnum.

custom_column_value_type_enum

Nested message and enum types in CustomColumnValueTypeEnum.

customer_status_enum

Nested message and enum types in CustomerStatusEnum.

data_driven_model_status_enum

Nested message and enum types in DataDrivenModelStatusEnum.

day_of_week_enum

Nested message and enum types in DayOfWeekEnum.

device_enum

Nested message and enum types in DeviceEnum.

floodlight_variable_data_type_enum

Nested message and enum types in FloodlightVariableDataTypeEnum.

floodlight_variable_type_enum

Nested message and enum types in FloodlightVariableTypeEnum.

gender_type_enum

Nested message and enum types in GenderTypeEnum.

geo_target_constant_status_enum

Nested message and enum types in GeoTargetConstantStatusEnum.

interaction_event_type_enum

Nested message and enum types in InteractionEventTypeEnum.

keyword_match_type_enum

Nested message and enum types in KeywordMatchTypeEnum.

label_status_enum

Nested message and enum types in LabelStatusEnum.

listing_group_filter_bidding_category_level_enum

Nested message and enum types in ListingGroupFilterBiddingCategoryLevelEnum.

listing_group_filter_custom_attribute_index_enum

Nested message and enum types in ListingGroupFilterCustomAttributeIndexEnum.

listing_group_filter_product_channel_enum

Nested message and enum types in ListingGroupFilterProductChannelEnum.

listing_group_filter_product_condition_enum

Nested message and enum types in ListingGroupFilterProductConditionEnum.

listing_group_filter_product_type_level_enum

Nested message and enum types in ListingGroupFilterProductTypeLevelEnum.

listing_group_filter_type_enum

Nested message and enum types in ListingGroupFilterTypeEnum.

listing_group_filter_vertical_enum

Nested message and enum types in ListingGroupFilterVerticalEnum.

listing_group_type_enum

Nested message and enum types in ListingGroupTypeEnum.

location_group_radius_units_enum

Nested message and enum types in LocationGroupRadiusUnitsEnum.

location_ownership_type_enum

Nested message and enum types in LocationOwnershipTypeEnum.

manager_link_status_enum

Nested message and enum types in ManagerLinkStatusEnum.

mime_type_enum

Nested message and enum types in MimeTypeEnum.

minute_of_hour_enum

Nested message and enum types in MinuteOfHourEnum.

mobile_app_vendor_enum

Nested message and enum types in MobileAppVendorEnum.

negative_geo_target_type_enum

Nested message and enum types in NegativeGeoTargetTypeEnum.

optimization_goal_type_enum

Nested message and enum types in OptimizationGoalTypeEnum.

positive_geo_target_type_enum

Nested message and enum types in PositiveGeoTargetTypeEnum.

product_bidding_category_level_enum

Nested message and enum types in ProductBiddingCategoryLevelEnum.

product_bidding_category_status_enum

Nested message and enum types in ProductBiddingCategoryStatusEnum.

product_channel_enum

Nested message and enum types in ProductChannelEnum.

product_channel_exclusivity_enum

Nested message and enum types in ProductChannelExclusivityEnum.

product_condition_enum

Nested message and enum types in ProductConditionEnum.

quality_score_bucket_enum

Nested message and enum types in QualityScoreBucketEnum.

search_ads360_field_category_enum

Nested message and enum types in SearchAds360FieldCategoryEnum.

search_ads360_field_data_type_enum

Nested message and enum types in SearchAds360FieldDataTypeEnum.

served_asset_field_type_enum

Nested message and enum types in ServedAssetFieldTypeEnum.

summary_row_setting_enum

Nested message and enum types in SummaryRowSettingEnum.

target_impression_share_location_enum

Nested message and enum types in TargetImpressionShareLocationEnum.

targeting_dimension_enum

Nested message and enum types in TargetingDimensionEnum.

user_list_type_enum

Nested message and enum types in UserListTypeEnum.

webpage_condition_operand_enum

Nested message and enum types in WebpageConditionOperandEnum.

webpage_condition_operator_enum

Nested message and enum types in WebpageConditionOperatorEnum.

Structs

AccountStatusEnum

Container for enum describing possible statuses of an account.

AccountTypeEnum

Container for enum describing engine account types.

AdGroupAdEngineStatusEnum

Container for enum describing possible AdGroupAd engine statuses.

AdGroupAdRotationModeEnum

Container for enum describing possible ad rotation modes of ads within an ad group.

AdGroupAdStatusEnum

Container for enum describing possible statuses of an AdGroupAd.

AdGroupCriterionEngineStatusEnum

Container for enum describing possible AdGroupCriterion engine statuses.

AdGroupCriterionStatusEnum

Message describing AdGroupCriterion statuses.

AdGroupEngineStatusEnum

Container for enum describing possible AdGroup engine statuses.

AdGroupStatusEnum

Container for enum describing possible statuses of an ad group.

AdGroupTypeEnum

Defines types of an ad group, specific to a particular campaign channel type. This type drives validations that restrict which entities can be added to the ad group.

AdNetworkTypeEnum

Container for enumeration of Google Ads network types.

AdServingOptimizationStatusEnum

Possible ad serving statuses of a campaign.

AdStrengthEnum

Container for enum describing possible ad strengths.

AdTypeEnum

Container for enum describing possible types of an ad.

AdvertisingChannelSubTypeEnum

An immutable specialization of an Advertising Channel.

AdvertisingChannelTypeEnum

The channel type a campaign may target to serve on.

AgeRangeTypeEnum

Container for enum describing the type of demographic age ranges.

AssetEngineStatusEnum

Container for enum describing possible Asset engine statuses.

AssetFieldTypeEnum

Container for enum describing the possible placements of an asset.

AssetGroupStatusEnum

Container for enum describing possible statuses of an asset group.

AssetLinkStatusEnum

Container for enum describing possible statuses of an asset link.

AssetSetAssetStatusEnum

Container for enum describing possible statuses of an asset set asset.

AssetSetLinkStatusEnum

Container for enum describing possible statuses of the linkage between asset set and its container.

AssetStatusEnum

Container for enum describing possible statuses of an asset.

AssetTypeEnum

Container for enum describing the types of asset.

AttributionModelEnum

Container for enum representing the attribution model that describes how to distribute credit for a particular conversion across potentially many prior interactions.

AttributionTypeEnum

Container for enum describing conversion attribution type in SearchAds360

BiddingStrategyStatusEnum

Message describing BiddingStrategy statuses.

BiddingStrategySystemStatusEnum

Message describing BiddingStrategy system statuses.

BiddingStrategyTypeEnum

Container for enum describing possible bidding strategy types.

BudgetDeliveryMethodEnum

Message describing Budget delivery methods. A delivery method determines the rate at which the Budget is spent.

BudgetPeriodEnum

Message describing Budget period.

CallConversionReportingStateEnum

Container for enum describing possible data types for call conversion reporting state.

CallToActionTypeEnum

Container for enum describing the call to action types.

CampaignCriterionStatusEnum

Message describing CampaignCriterion statuses.

CampaignServingStatusEnum

Message describing Campaign serving statuses.

CampaignStatusEnum

Container for enum describing possible statuses of a campaign.

ConversionActionCategoryEnum

Container for enum describing the category of conversions that are associated with a ConversionAction.

ConversionActionStatusEnum

Container for enum describing possible statuses of a conversion action.

ConversionActionTypeEnum

Container for enum describing possible types of a conversion action.

ConversionCustomVariableCardinalityEnum

Container for enum describing the family of a conversion custom variable.

ConversionCustomVariableFamilyEnum

Container for enum describing the family of a conversion custom variable.

ConversionCustomVariableStatusEnum

Container for enum describing possible statuses of a conversion custom variable.

ConversionStatusEnum

Container for enum describing possible statuses of a conversion.

ConversionTrackingStatusEnum

Container for enum representing the conversion tracking status of the customer.

CriterionTypeEnum

The possible types of a criterion.

CustomColumnRenderTypeEnum

The render type of custom columns.

CustomColumnValueTypeEnum

The value type of custom columns.

CustomerStatusEnum

Container for enum describing possible statuses of a customer.

DataDrivenModelStatusEnum

Container for enum indicating data driven model status.

DayOfWeekEnum

Container for enumeration of days of the week, for example, “Monday”.

DeviceEnum

Container for enumeration of Google Ads devices available for targeting.

FloodlightVariableDataTypeEnum

Container for enum describing Floodlight variable type defined in Search Ads 360.

FloodlightVariableTypeEnum

Container for enum describing Floodlight variable type defined in Search Ads 360.

GenderTypeEnum

Container for enum describing the type of demographic genders.

GeoTargetConstantStatusEnum

Container for describing the status of a geo target constant.

InteractionEventTypeEnum

Container for enum describing types of payable and free interactions.

KeywordMatchTypeEnum

Message describing Keyword match types.

LabelStatusEnum

Container for enum describing possible status of a label.

ListingGroupFilterBiddingCategoryLevelEnum

Container for enum describing the levels of bidding category used in ListingGroupFilterDimension.

ListingGroupFilterCustomAttributeIndexEnum

Container for enum describing the indexes of custom attribute used in ListingGroupFilterDimension.

ListingGroupFilterProductChannelEnum

Locality of a product offer.

ListingGroupFilterProductConditionEnum

Condition of a product offer.

ListingGroupFilterProductTypeLevelEnum

Level of the type of a product offer.

ListingGroupFilterTypeEnum

Container for enum describing the type of the listing group filter node.

ListingGroupFilterVerticalEnum

Container for enum describing the type of the vertical a listing group filter tree represents.

ListingGroupTypeEnum

Container for enum describing the type of the listing group.

LocationGroupRadiusUnitsEnum

Container for enum describing unit of radius in location group.

LocationOwnershipTypeEnum

Container for enum describing possible types of a location ownership.

ManagerLinkStatusEnum

Container for enum describing possible status of a manager and client link.

MimeTypeEnum

Container for enum describing the mime types.

MinuteOfHourEnum

Container for enumeration of quarter-hours.

MobileAppVendorEnum

Container for enum describing different types of mobile app vendors.

NegativeGeoTargetTypeEnum

Container for enum describing possible negative geo target types.

OptimizationGoalTypeEnum

Container for enum describing the type of optimization goal.

PositiveGeoTargetTypeEnum

Container for enum describing possible positive geo target types.

ProductBiddingCategoryLevelEnum

Level of a product bidding category.

ProductBiddingCategoryStatusEnum

Status of the product bidding category.

ProductChannelEnum

Locality of a product offer.

ProductChannelExclusivityEnum

Availability of a product offer.

ProductConditionEnum

Condition of a product offer.

QualityScoreBucketEnum

The relative performance compared to other advertisers.

SearchAds360FieldCategoryEnum

Container for enum that determines if the described artifact is a resource or a field, and if it is a field, when it segments search queries.

SearchAds360FieldDataTypeEnum

Container holding the various data types.

ServedAssetFieldTypeEnum

Container for enum describing possible asset field types.

SummaryRowSettingEnum

Indicates summary row setting in request parameter.

TargetImpressionShareLocationEnum

Container for enum describing where on the first search results page the automated bidding system should target impressions for the TargetImpressionShare bidding strategy.

TargetingDimensionEnum

The dimensions that can be targeted.

UserListTypeEnum

The user list types.

WebpageConditionOperandEnum

Container for enum describing webpage condition operand in webpage criterion.

WebpageConditionOperatorEnum

Container for enum describing webpage condition operator in webpage criterion.