Module google_api_proto::google::ads::searchads360::v0::errors

Modules

authentication_error_enum

Nested message and enum types in AuthenticationErrorEnum.

authorization_error_enum

Nested message and enum types in AuthorizationErrorEnum.

custom_column_error_enum

Nested message and enum types in CustomColumnErrorEnum.

date_error_enum

Nested message and enum types in DateErrorEnum.

date_range_error_enum

Nested message and enum types in DateRangeErrorEnum.

distinct_error_enum

Nested message and enum types in DistinctErrorEnum.

error_code

Nested message and enum types in ErrorCode.

error_location

Nested message and enum types in ErrorLocation.

header_error_enum

Nested message and enum types in HeaderErrorEnum.

internal_error_enum

Nested message and enum types in InternalErrorEnum.

invalid_parameter_error_enum

Nested message and enum types in InvalidParameterErrorEnum.

query_error_enum

Nested message and enum types in QueryErrorEnum.

quota_error_details

Nested message and enum types in QuotaErrorDetails.

quota_error_enum

Nested message and enum types in QuotaErrorEnum.

request_error_enum

Nested message and enum types in RequestErrorEnum.

size_limit_error_enum

Nested message and enum types in SizeLimitErrorEnum.

Structs

AuthenticationErrorEnum

Container for enum describing possible authentication errors.

AuthorizationErrorEnum

Container for enum describing possible authorization errors.

CustomColumnErrorEnum

Container for enum describing possible custom column errors.

DateErrorEnum

Container for enum describing possible date errors.

DateRangeErrorEnum

Container for enum describing possible date range errors.

DistinctErrorEnum

Container for enum describing possible distinct errors.

ErrorCode

The error reason represented by type and enum.

ErrorDetails

Additional error details.

ErrorLocation

Describes the part of the request proto that caused the error.

HeaderErrorEnum

Container for enum describing possible header errors.

InternalErrorEnum

Container for enum describing possible internal errors.

InvalidParameterErrorEnum

Container for enum describing possible errors from an invalid parameter.

QueryErrorEnum

Container for enum describing possible query errors.

QuotaErrorDetails

Additional quota error details when there is QuotaError.

QuotaErrorEnum

Container for enum describing possible quota errors.

RequestErrorEnum

Container for enum describing possible request errors.

SearchAds360Error

SearchAds360-specific error.

SearchAds360Failure

Describes how a Search Ads 360 API call failed. It’s returned inside google.rpc.Status.details when a call fails.

SizeLimitErrorEnum

Container for enum describing possible size limit errors.