Enum google_api_proto::google::ads::searchads360::v0::errors::request_error_enum::RequestError

#[repr(i32)]
pub enum RequestError {

    Unspecified = 0,
    Unknown = 1,
    ResourceNameMissing = 3,
    ResourceNameMalformed = 4,
    BadResourceId = 17,
    InvalidProductName = 35,
    InvalidCustomerId = 16,
    OperationRequired = 5,
    ResourceNotFound = 6,
    InvalidPageToken = 7,
    ExpiredPageToken = 8,
    InvalidPageSize = 22,
    RequiredFieldMissing = 9,
    ImmutableField = 11,
    TooManyMutateOperations = 13,
    CannotBeExecutedByManagerAccount = 14,
    CannotModifyForeignField = 15,
    InvalidEnumValue = 18,
    LoginCustomerIdParameterMissing = 20,
    LoginOrLinkedCustomerIdParameterRequired = 34,
    ValidateOnlyRequestHasPageToken = 21,
    CannotReturnSummaryRowForRequestWithoutMetrics = 29,
    CannotReturnSummaryRowForValidateOnlyRequests = 30,
    InconsistentReturnSummaryRowValue = 31,
    TotalResultsCountNotOriginallyRequested = 32,
    RpcDeadlineTooShort = 33,
    ProductNotSupported = 37,

}

Enum describing possible request errors.

Variants

Unspecified = 0

Enum unspecified.

Unknown = 1

The received error code is not known in this version.

ResourceNameMissing = 3

Resource name is required for this request.

ResourceNameMalformed = 4

Resource name provided is malformed.

BadResourceId = 17

Resource name provided is malformed.

InvalidProductName = 35

Product name is invalid.

InvalidCustomerId = 16

Customer ID is invalid.

OperationRequired = 5

Mutate operation should have either create, update, or remove specified.

ResourceNotFound = 6

Requested resource not found.

InvalidPageToken = 7

Next page token specified in user request is invalid.

ExpiredPageToken = 8

Next page token specified in user request has expired.

InvalidPageSize = 22

Page size specified in user request is invalid.

RequiredFieldMissing = 9

Required field is missing.

ImmutableField = 11

The field cannot be modified because it’s immutable. It’s also possible that the field can be modified using ‘create’ operation but not ‘update’.

TooManyMutateOperations = 13

Received too many entries in request.

CannotBeExecutedByManagerAccount = 14

Request cannot be executed by a manager account.

CannotModifyForeignField = 15

Mutate request was attempting to modify a readonly field. For instance, Budget fields can be requested for Ad Group, but are read-only for adGroups:mutate.

InvalidEnumValue = 18

Enum value is not permitted.

LoginCustomerIdParameterMissing = 20

The login-customer-id parameter is required for this request.

LoginOrLinkedCustomerIdParameterRequired = 34

Either login-customer-id or linked-customer-id parameter is required for this request.

ValidateOnlyRequestHasPageToken = 21

page_token is set in the validate only request

CannotReturnSummaryRowForRequestWithoutMetrics = 29

return_summary_row cannot be enabled if request did not select any metrics field.

CannotReturnSummaryRowForValidateOnlyRequests = 30

return_summary_row should not be enabled for validate only requests.

InconsistentReturnSummaryRowValue = 31

return_summary_row parameter value should be the same between requests with page_token field set and their original request.

TotalResultsCountNotOriginallyRequested = 32

The total results count cannot be returned if it was not requested in the original request.

RpcDeadlineTooShort = 33

Deadline specified by the client was too short.

ProductNotSupported = 37

The product associated with the request is not supported for the current request.

Implementations

impl RequestError

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of RequestError.

pub fn from_i32(value: i32) -> Option<RequestError>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a RequestError, or None if value is not a valid variant.

impl RequestError

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.