Module google_api_proto::google::ads::searchads360::v0::resources

Modules

accessible_bidding_strategy

Nested message and enum types in AccessibleBiddingStrategy.

ad

Nested message and enum types in Ad.

ad_group_bid_modifier

Nested message and enum types in AdGroupBidModifier.

ad_group_criterion

Nested message and enum types in AdGroupCriterion.

asset

Nested message and enum types in Asset.

asset_group_signal

Nested message and enum types in AssetGroupSignal.

bidding_strategy

Nested message and enum types in BiddingStrategy.

campaign

Nested message and enum types in Campaign.

campaign_criterion

Nested message and enum types in CampaignCriterion.

conversion_action

Nested message and enum types in ConversionAction.

conversion_custom_variable

Nested message and enum types in ConversionCustomVariable.

listing_group_filter_dimension

Nested message and enum types in ListingGroupFilterDimension.

Structs

AccessibleBiddingStrategy

Represents a view of BiddingStrategies owned by and shared with the customer.

Ad

An ad.

AdGroup

An ad group.

AdGroupAd

An ad group ad.

AdGroupAdLabel

A relationship between an ad group ad and a label.

AdGroupAsset

A link between an ad group and an asset.

AdGroupAssetSet

AdGroupAssetSet is the linkage between an ad group and an asset set. Creating an AdGroupAssetSet links an asset set with an ad group.

AdGroupAudienceView

An ad group audience view. Includes performance data from interests and remarketing lists for Display Network and YouTube Network ads, and remarketing lists for search ads (RLSA), aggregated at the audience level.

AdGroupBidModifier

Represents an ad group bid modifier.

AdGroupCriterion

An ad group criterion. The ad_group_criterion report only returns criteria that were explicitly added to the ad group.

AdGroupCriterionLabel

A relationship between an ad group criterion and a label.

AdGroupLabel

A relationship between an ad group and a label.

AgeRangeView

An age range view.

Asset

Asset is a part of an ad which can be shared across multiple ads. It can be an image (ImageAsset), a video (YoutubeVideoAsset), etc. Assets are immutable and cannot be removed. To stop an asset from serving, remove the asset from the entity that is using it.

AssetGroup

An asset group. AssetGroupAsset is used to link an asset to the asset group. AssetGroupSignal is used to associate a signal to an asset group.

AssetGroupAsset

AssetGroupAsset is the link between an asset and an asset group. Adding an AssetGroupAsset links an asset with an asset group.

AssetGroupAssetCombinationData

Asset group asset combination data

AssetGroupListingGroupFilter

AssetGroupListingGroupFilter represents a listing group filter tree node in an asset group.

AssetGroupSignal

AssetGroupSignal represents a signal in an asset group. The existence of a signal tells the performance max campaign who’s most likely to convert. Performance Max uses the signal to look for new people with similar or stronger intent to find conversions across Search, Display, Video, and more.

AssetGroupTopCombinationView

A view on the usage of ad group ad asset combination.

AssetSet

An asset set representing a collection of assets. Use AssetSetAsset to link an asset to the asset set.

AssetSetAsset

AssetSetAsset is the link between an asset and an asset set. Adding an AssetSetAsset links an asset with an asset set.

Audience

Audience is an effective targeting option that lets you intersect different segment attributes, such as detailed demographics and affinities, to create audiences that represent sections of your target segments.

BiddingStrategy

A bidding strategy.

Campaign

A campaign.

CampaignAsset

A link between a Campaign and an Asset.

CampaignAssetSet

CampaignAssetSet is the linkage between a campaign and an asset set. Adding a CampaignAssetSet links an asset set with a campaign.

CampaignAudienceView

A campaign audience view. Includes performance data from interests and remarketing lists for Display Network and YouTube Network ads, and remarketing lists for search ads (RLSA), aggregated by campaign and audience criterion. This view only includes audiences attached at the campaign level.

CampaignBudget

A campaign budget.

CampaignCriterion

A campaign criterion.

CampaignLabel

Represents a relationship between a campaign and a label.

CartDataSalesView

Cart data sales view.

Conversion

A conversion.

ConversionAction

A conversion action.

ConversionCustomVariable

A conversion custom variable. See “About custom Floodlight metrics and dimensions in the new Search Ads 360” at https://support.google.com/sa360/answer/13567857

ConversionTrackingSetting

A collection of customer-wide settings related to Search Ads 360 Conversion Tracking.

CustomColumn

A custom column. See Search Ads 360 custom column at https://support.google.com/sa360/answer/9633916

Customer

A customer.

CustomerAsset

A link between a customer and an asset.

CustomerAssetSet

CustomerAssetSet is the linkage between a customer and an asset set. Adding a CustomerAssetSet links an asset set with a customer.

CustomerClient

A link between the given customer and a client customer. CustomerClients only exist for manager customers. All direct and indirect client customers are included, as well as the manager itself.

CustomerManagerLink

Represents customer-manager link relationship.

DoubleClickCampaignManagerSetting

DoubleClick Campaign Manager (DCM) setting for a manager customer.

DynamicSearchAdsSearchTermView

A dynamic search ads search term view.

GenderView

A gender view. The gender_view resource reflects the effective serving state, rather than what criteria were added. An ad group without gender criteria by default shows to all genders, so all genders appear in gender_view with stats.

GeoTargetConstant

A geo target constant.

KeywordView

A keyword view.

Label

A label.

LanguageConstant

A language.

ListingGroupFilterDimension

Listing dimensions for the asset group listing group filter.

ListingGroupFilterDimensionPath

The path defining of dimensions defining a listing group filter.

LocationView

A location view summarizes the performance of campaigns by a Location criterion.

ProductBiddingCategoryConstant

A Product Bidding Category.

ProductGroupView

A product group view.

SearchAds360Field

A field or resource (artifact) used by SearchAds360Service.

ShoppingPerformanceView

Shopping performance view. Provides Shopping campaign statistics aggregated at several product dimension levels. Product dimension values from Merchant Center such as brand, category, custom attributes, product condition and product type will reflect the state of each dimension as of the date and time when the corresponding event was recorded.

UserList

A user list. This is a list of users a customer may target.

Visit

A visit.

WebpageView

A webpage view.