Struct google_api_proto::google::ads::searchads360::v0::resources::Conversion

pub struct Conversion {

    pub resource_name: String,
    pub id: Option<i64>,
    pub criterion_id: Option<i64>,
    pub merchant_id: Option<i64>,
    pub ad_id: Option<i64>,
    pub click_id: Option<String>,
    pub visit_id: Option<i64>,
    pub advertiser_conversion_id: Option<String>,
    pub product_id: Option<String>,
    pub product_channel: Option<i32>,
    pub product_language_code: Option<String>,
    pub product_store_id: Option<String>,
    pub product_country_code: Option<String>,
    pub attribution_type: Option<i32>,
    pub conversion_date_time: Option<String>,
    pub conversion_last_modified_date_time: Option<String>,
    pub conversion_visit_date_time: Option<String>,
    pub conversion_quantity: Option<i64>,
    pub conversion_revenue_micros: Option<i64>,
    pub floodlight_original_revenue: Option<i64>,
    pub floodlight_order_id: Option<String>,
    pub status: Option<i32>,
    pub asset_id: Option<i64>,
    pub asset_field_type: Option<i32>,

}

A conversion.

Fields
resource_name: String

Output only. The resource name of the conversion. Conversion resource names have the form:

customers/{customer_id}/conversions/{ad_group_id}~{criterion_id}~{ds_conversion_id}

id: Option<i64>

Output only. The ID of the conversion

criterion_id: Option<i64>

Output only. Search Ads 360 criterion ID. A value of 0 indicates that the criterion is unattributed.

merchant_id: Option<i64>

Output only. The Search Ads 360 inventory account ID containing the product that was clicked on. Search Ads 360 generates this ID when you link an inventory account in Search Ads 360.

ad_id: Option<i64>

Output only. Ad ID. A value of 0 indicates that the ad is unattributed.

click_id: Option<String>

Output only. A unique string, for the visit that the conversion is attributed to, that is passed to the landing page as the click id URL parameter.

visit_id: Option<i64>

Output only. The Search Ads 360 visit ID that the conversion is attributed to.

advertiser_conversion_id: Option<String>

Output only. For offline conversions, this is an ID provided by advertisers. If an advertiser doesn’t specify such an ID, Search Ads 360 generates one. For online conversions, this is equal to the id column or the floodlight_order_id column depending on the advertiser’s Floodlight instructions.

product_id: Option<String>

Output only. The ID of the product clicked on.

product_channel: Option<i32>

Output only. The sales channel of the product that was clicked on: Online or Local.

product_language_code: Option<String>

Output only. The language (ISO-639-1) that has been set for the Merchant Center feed containing data about the product.

product_store_id: Option<String>

Output only. The store in the Local Inventory Ad that was clicked on. This should match the store IDs used in your local products feed.

product_country_code: Option<String>

Output only. The country (ISO-3166-format) registered for the inventory feed that contains the product clicked on.

attribution_type: Option<i32>

Output only. What the conversion is attributed to: Visit or Keyword+Ad.

conversion_date_time: Option<String>

Output only. The timestamp of the conversion event.

conversion_last_modified_date_time: Option<String>

Output only. The timestamp of the last time the conversion was modified.

conversion_visit_date_time: Option<String>

Output only. The timestamp of the visit that the conversion is attributed to.

conversion_quantity: Option<i64>

Output only. The quantity of items recorded by the conversion, as determined by the qty url parameter. The advertiser is responsible for dynamically populating the parameter (such as number of items sold in the conversion), otherwise it defaults to 1.

conversion_revenue_micros: Option<i64>

Output only. The adjusted revenue in micros for the conversion event. This will always be in the currency of the serving account.

floodlight_original_revenue: Option<i64>

Output only. The original, unchanged revenue associated with the Floodlight event (in the currency of the current report), before Floodlight currency instruction modifications.

floodlight_order_id: Option<String>

Output only. The Floodlight order ID provided by the advertiser for the conversion.

status: Option<i32>

Output only. The status of the conversion, either ENABLED or REMOVED..

asset_id: Option<i64>

Output only. ID of the asset which was interacted with during the conversion event.

asset_field_type: Option<i32>

Output only. Asset field type of the conversion event.

Implementations

impl Conversion

pub fn id(&self) -> i64

Returns the value of id, or the default value if id is unset.

pub fn criterion_id(&self) -> i64

Returns the value of criterion_id, or the default value if criterion_id is unset.

pub fn merchant_id(&self) -> i64

Returns the value of merchant_id, or the default value if merchant_id is unset.

pub fn ad_id(&self) -> i64

Returns the value of ad_id, or the default value if ad_id is unset.

pub fn click_id(&self) -> &str

Returns the value of click_id, or the default value if click_id is unset.

pub fn visit_id(&self) -> i64

Returns the value of visit_id, or the default value if visit_id is unset.

pub fn advertiser_conversion_id(&self) -> &str

Returns the value of advertiser_conversion_id, or the default value if advertiser_conversion_id is unset.

pub fn product_id(&self) -> &str

Returns the value of product_id, or the default value if product_id is unset.

pub fn product_channel(&self) -> ProductChannel

Returns the enum value of product_channel, or the default if the field is unset or set to an invalid enum value.

pub fn set_product_channel(&mut self, value: ProductChannel)

Sets product_channel to the provided enum value.

pub fn product_language_code(&self) -> &str

Returns the value of product_language_code, or the default value if product_language_code is unset.

pub fn product_store_id(&self) -> &str

Returns the value of product_store_id, or the default value if product_store_id is unset.

pub fn product_country_code(&self) -> &str

Returns the value of product_country_code, or the default value if product_country_code is unset.

pub fn attribution_type(&self) -> AttributionType

Returns the enum value of attribution_type, or the default if the field is unset or set to an invalid enum value.

pub fn set_attribution_type(&mut self, value: AttributionType)

Sets attribution_type to the provided enum value.

pub fn conversion_date_time(&self) -> &str

Returns the value of conversion_date_time, or the default value if conversion_date_time is unset.

pub fn conversion_last_modified_date_time(&self) -> &str

Returns the value of conversion_last_modified_date_time, or the default value if conversion_last_modified_date_time is unset.

pub fn conversion_visit_date_time(&self) -> &str

Returns the value of conversion_visit_date_time, or the default value if conversion_visit_date_time is unset.

pub fn conversion_quantity(&self) -> i64

Returns the value of conversion_quantity, or the default value if conversion_quantity is unset.

pub fn conversion_revenue_micros(&self) -> i64

Returns the value of conversion_revenue_micros, or the default value if conversion_revenue_micros is unset.

pub fn floodlight_original_revenue(&self) -> i64

Returns the value of floodlight_original_revenue, or the default value if floodlight_original_revenue is unset.

pub fn floodlight_order_id(&self) -> &str

Returns the value of floodlight_order_id, or the default value if floodlight_order_id is unset.

pub fn status(&self) -> ConversionStatus

Returns the enum value of status, or the default if the field is unset or set to an invalid enum value.

pub fn set_status(&mut self, value: ConversionStatus)

Sets status to the provided enum value.

pub fn asset_id(&self) -> i64

Returns the value of asset_id, or the default value if asset_id is unset.

pub fn asset_field_type(&self) -> AssetFieldType

Returns the enum value of asset_field_type, or the default if the field is unset or set to an invalid enum value.

pub fn set_asset_field_type(&mut self, value: AssetFieldType)

Sets asset_field_type to the provided enum value.