Struct google_api_proto::google::ads::searchads360::v0::services::SearchAds360Row

pub struct SearchAds360Row {

    pub ad_group: Option<AdGroup>,
    pub ad_group_ad: Option<AdGroupAd>,
    pub ad_group_ad_label: Option<AdGroupAdLabel>,
    pub ad_group_asset: Option<AdGroupAsset>,
    pub ad_group_asset_set: Option<AdGroupAssetSet>,
    pub ad_group_audience_view: Option<AdGroupAudienceView>,
    pub ad_group_bid_modifier: Option<AdGroupBidModifier>,
    pub ad_group_criterion: Option<AdGroupCriterion>,
    pub ad_group_criterion_label: Option<AdGroupCriterionLabel>,
    pub ad_group_label: Option<AdGroupLabel>,
    pub age_range_view: Option<AgeRangeView>,
    pub asset: Option<Asset>,
    pub asset_group_asset: Option<AssetGroupAsset>,
    pub asset_group_signal: Option<AssetGroupSignal>,
    pub asset_group_listing_group_filter: Option<AssetGroupListingGroupFilter>,
    pub asset_group_top_combination_view: Option<AssetGroupTopCombinationView>,
    pub asset_group: Option<AssetGroup>,
    pub asset_set_asset: Option<AssetSetAsset>,
    pub asset_set: Option<AssetSet>,
    pub bidding_strategy: Option<BiddingStrategy>,
    pub campaign_budget: Option<CampaignBudget>,
    pub campaign: Option<Campaign>,
    pub campaign_asset: Option<CampaignAsset>,
    pub campaign_asset_set: Option<CampaignAssetSet>,
    pub campaign_audience_view: Option<CampaignAudienceView>,
    pub campaign_criterion: Option<CampaignCriterion>,
    pub campaign_label: Option<CampaignLabel>,
    pub cart_data_sales_view: Option<CartDataSalesView>,
    pub audience: Option<Audience>,
    pub conversion_action: Option<ConversionAction>,
    pub conversion_custom_variable: Option<ConversionCustomVariable>,
    pub customer: Option<Customer>,
    pub customer_asset: Option<CustomerAsset>,
    pub customer_asset_set: Option<CustomerAssetSet>,
    pub accessible_bidding_strategy: Option<AccessibleBiddingStrategy>,
    pub customer_manager_link: Option<CustomerManagerLink>,
    pub customer_client: Option<CustomerClient>,
    pub dynamic_search_ads_search_term_view: Option<DynamicSearchAdsSearchTermView>,
    pub gender_view: Option<GenderView>,
    pub geo_target_constant: Option<GeoTargetConstant>,
    pub keyword_view: Option<KeywordView>,
    pub label: Option<Label>,
    pub language_constant: Option<LanguageConstant>,
    pub location_view: Option<LocationView>,
    pub product_bidding_category_constant: Option<ProductBiddingCategoryConstant>,
    pub product_group_view: Option<ProductGroupView>,
    pub shopping_performance_view: Option<ShoppingPerformanceView>,
    pub user_list: Option<UserList>,
    pub webpage_view: Option<WebpageView>,
    pub visit: Option<Visit>,
    pub conversion: Option<Conversion>,
    pub metrics: Option<Metrics>,
    pub segments: Option<Segments>,
    pub custom_columns: Vec<Value>,

}

A returned row from the query.

Fields
ad_group: Option<AdGroup>

The ad group referenced in the query.

ad_group_ad: Option<AdGroupAd>

The ad referenced in the query.

ad_group_ad_label: Option<AdGroupAdLabel>

The ad group ad label referenced in the query.

ad_group_asset: Option<AdGroupAsset>

The ad group asset referenced in the query.

ad_group_asset_set: Option<AdGroupAssetSet>

The ad group asset set referenced in the query.

ad_group_audience_view: Option<AdGroupAudienceView>

The ad group audience view referenced in the query.

ad_group_bid_modifier: Option<AdGroupBidModifier>

The bid modifier referenced in the query.

ad_group_criterion: Option<AdGroupCriterion>

The criterion referenced in the query.

ad_group_criterion_label: Option<AdGroupCriterionLabel>

The ad group criterion label referenced in the query.

ad_group_label: Option<AdGroupLabel>

The ad group label referenced in the query.

age_range_view: Option<AgeRangeView>

The age range view referenced in the query.

asset: Option<Asset>

The asset referenced in the query.

asset_group_asset: Option<AssetGroupAsset>

The asset group asset referenced in the query.

asset_group_signal: Option<AssetGroupSignal>

The asset group signal referenced in the query.

asset_group_listing_group_filter: Option<AssetGroupListingGroupFilter>

The asset group listing group filter referenced in the query.

asset_group_top_combination_view: Option<AssetGroupTopCombinationView>

The asset group top combination view referenced in the query.

asset_group: Option<AssetGroup>

The asset group referenced in the query.

asset_set_asset: Option<AssetSetAsset>

The asset set asset referenced in the query.

asset_set: Option<AssetSet>

The asset set referenced in the query.

bidding_strategy: Option<BiddingStrategy>

The bidding strategy referenced in the query.

campaign_budget: Option<CampaignBudget>

The campaign budget referenced in the query.

campaign: Option<Campaign>

The campaign referenced in the query.

campaign_asset: Option<CampaignAsset>

The campaign asset referenced in the query.

campaign_asset_set: Option<CampaignAssetSet>

The campaign asset set referenced in the query.

campaign_audience_view: Option<CampaignAudienceView>

The campaign audience view referenced in the query.

campaign_criterion: Option<CampaignCriterion>

The campaign criterion referenced in the query.

campaign_label: Option<CampaignLabel>

The campaign label referenced in the query.

cart_data_sales_view: Option<CartDataSalesView>

The cart data sales view referenced in the query.

audience: Option<Audience>

The Audience referenced in the query.

conversion_action: Option<ConversionAction>

The conversion action referenced in the query.

conversion_custom_variable: Option<ConversionCustomVariable>

The conversion custom variable referenced in the query.

customer: Option<Customer>

The customer referenced in the query.

customer_asset: Option<CustomerAsset>

The customer asset referenced in the query.

customer_asset_set: Option<CustomerAssetSet>

The customer asset set referenced in the query.

accessible_bidding_strategy: Option<AccessibleBiddingStrategy>

The accessible bidding strategy referenced in the query.

customer_manager_link: Option<CustomerManagerLink>

The CustomerManagerLink referenced in the query.

customer_client: Option<CustomerClient>

The CustomerClient referenced in the query.

dynamic_search_ads_search_term_view: Option<DynamicSearchAdsSearchTermView>

The dynamic search ads search term view referenced in the query.

gender_view: Option<GenderView>

The gender view referenced in the query.

geo_target_constant: Option<GeoTargetConstant>

The geo target constant referenced in the query.

keyword_view: Option<KeywordView>

The keyword view referenced in the query.

label: Option<Label>

The label referenced in the query.

language_constant: Option<LanguageConstant>

The language constant referenced in the query.

location_view: Option<LocationView>

The location view referenced in the query.

product_bidding_category_constant: Option<ProductBiddingCategoryConstant>

The Product Bidding Category referenced in the query.

product_group_view: Option<ProductGroupView>

The product group view referenced in the query.

shopping_performance_view: Option<ShoppingPerformanceView>

The shopping performance view referenced in the query.

user_list: Option<UserList>

The user list referenced in the query.

webpage_view: Option<WebpageView>

The webpage view referenced in the query.

visit: Option<Visit>

The event level visit referenced in the query.

conversion: Option<Conversion>

The event level conversion referenced in the query.

metrics: Option<Metrics>

The metrics.

segments: Option<Segments>

The segments.

custom_columns: Vec<Value>

The custom columns.