Module google_api_proto::google::ai::generativelanguage::v1

Modules

candidate

Nested message and enum types in Candidate.

generate_content_response

Nested message and enum types in GenerateContentResponse.

generative_service_client

Generated client implementations.

model_service_client

Generated client implementations.

part

Nested message and enum types in Part.

safety_rating

Nested message and enum types in SafetyRating.

safety_setting

Nested message and enum types in SafetySetting.

Structs

BatchEmbedContentsRequest

Batch request to get embeddings from the model for a list of prompts.

BatchEmbedContentsResponse

The response to a BatchEmbedContentsRequest.

Blob

Raw media bytes.

Candidate

A response candidate generated from the model.

CitationMetadata

A collection of source attributions for a piece of content.

CitationSource

A citation to a source for a portion of a specific response.

Content

The base structured datatype containing multi-part content of a message.

ContentEmbedding

A list of floats representing an embedding.

CountTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountTokensResponse

A response from CountTokens.

EmbedContentRequest

Request containing the Content for the model to embed.

EmbedContentResponse

The response to an EmbedContentRequest.

GenerateContentRequest

Request to generate a completion from the model.

GenerateContentResponse

Response from the model supporting multiple candidate responses.

GenerationConfig

Configuration options for model generation and outputs. Not all parameters are configurable for every model.

GetModelRequest

Request for getting information about a specific Model.

ListModelsRequest

Request for listing all Models.

ListModelsResponse

Response from ListModel containing a paginated list of Models.

Model

Information about a Generative Language Model.

Part

A datatype containing media that is part of a multi-part Content message.

SafetyRating

Safety rating for a piece of content.

SafetySetting

Safety setting, affecting the safety-blocking behavior.

Enums

HarmCategory

The category of a rating.

TaskType

Type of task for which the embedding will be used.