Enum google_api_proto::google::analytics::admin::v1alpha::ChangeHistoryResourceType

#[repr(i32)]
pub enum ChangeHistoryResourceType {

    Unspecified = 0,
    Account = 1,
    Property = 2,
    FirebaseLink = 6,
    GoogleAdsLink = 7,
    GoogleSignalsSettings = 8,
    ConversionEvent = 9,
    MeasurementProtocolSecret = 10,
    CustomDimension = 11,
    CustomMetric = 12,
    DataRetentionSettings = 13,
    DisplayVideo360AdvertiserLink = 14,
    DisplayVideo360AdvertiserLinkProposal = 15,
    SearchAds360Link = 16,
    DataStream = 18,
    AttributionSettings = 20,
    ExpandedDataSet = 21,
    ChannelGroup = 22,
    EnhancedMeasurementSettings = 24,
    DataRedactionSettings = 25,
    SkadnetworkConversionValueSchema = 26,
    AdsenseLink = 27,
    Audience = 28,
    EventCreateRule = 29,
    CalculatedMetric = 31,

}

Types of resources whose changes may be returned from change history.

Variants

Unspecified = 0

Resource type unknown or not specified.

Account = 1

Account resource

Property = 2

Property resource

FirebaseLink = 6

FirebaseLink resource

GoogleAdsLink = 7

GoogleAdsLink resource

GoogleSignalsSettings = 8

GoogleSignalsSettings resource

ConversionEvent = 9

ConversionEvent resource

MeasurementProtocolSecret = 10

MeasurementProtocolSecret resource

CustomDimension = 11

CustomDimension resource

CustomMetric = 12

CustomMetric resource

DataRetentionSettings = 13

DataRetentionSettings resource

DisplayVideo360AdvertiserLink = 14

DisplayVideo360AdvertiserLink resource

DisplayVideo360AdvertiserLinkProposal = 15

DisplayVideo360AdvertiserLinkProposal resource

SearchAds360Link = 16

SearchAds360Link resource

DataStream = 18

DataStream resource

AttributionSettings = 20

AttributionSettings resource

ExpandedDataSet = 21

ExpandedDataSet resource

ChannelGroup = 22

ChannelGroup resource

EnhancedMeasurementSettings = 24

EnhancedMeasurementSettings resource

DataRedactionSettings = 25

DataRedactionSettings resource

SkadnetworkConversionValueSchema = 26

SKAdNetworkConversionValueSchema resource

AdsenseLink = 27

AdSenseLink resource

Audience = 28

Audience resource

EventCreateRule = 29

EventCreateRule resource

CalculatedMetric = 31

CalculatedMetric resource

Implementations

impl ChangeHistoryResourceType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ChangeHistoryResourceType.

pub fn from_i32(value: i32) -> Option<ChangeHistoryResourceType>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ChangeHistoryResourceType, or None if value is not a valid variant.

impl ChangeHistoryResourceType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.