Module google_api_proto::google::analytics::admin::v1beta

Modules

access_filter

Nested message and enum types in AccessFilter.

access_filter_expression

Nested message and enum types in AccessFilterExpression.

access_numeric_filter

Nested message and enum types in AccessNumericFilter.

access_order_by

Nested message and enum types in AccessOrderBy.

access_string_filter

Nested message and enum types in AccessStringFilter.

analytics_admin_service_client

Generated client implementations.

change_history_change

Nested message and enum types in ChangeHistoryChange.

conversion_event

Nested message and enum types in ConversionEvent.

custom_dimension

Nested message and enum types in CustomDimension.

custom_metric

Nested message and enum types in CustomMetric.

data_retention_settings

Nested message and enum types in DataRetentionSettings.

data_stream

Nested message and enum types in DataStream.

numeric_value

Nested message and enum types in NumericValue.

Structs

AccessBetweenFilter

To express that the result needs to be between two numbers (inclusive).

AccessDateRange

A contiguous range of days: startDate, startDate + 1, …, endDate.

AccessDimension

Dimensions are attributes of your data. For example, the dimension userEmail indicates the email of the user that accessed reporting data. Dimension values in report responses are strings.

AccessDimensionHeader

Describes a dimension column in the report. Dimensions requested in a report produce column entries within rows and DimensionHeaders. However, dimensions used exclusively within filters or expressions do not produce columns in a report; correspondingly, those dimensions do not produce headers.

AccessDimensionValue

The value of a dimension.

AccessFilter

An expression to filter dimension or metric values.

AccessFilterExpression

Expresses dimension or metric filters. The fields in the same expression need to be either all dimensions or all metrics.

AccessFilterExpressionList

A list of filter expressions.

AccessInListFilter

The result needs to be in a list of string values.

AccessMetric

The quantitative measurements of a report. For example, the metric accessCount is the total number of data access records.

AccessMetricHeader

Describes a metric column in the report. Visible metrics requested in a report produce column entries within rows and MetricHeaders. However, metrics used exclusively within filters or expressions do not produce columns in a report; correspondingly, those metrics do not produce headers.

AccessMetricValue

The value of a metric.

AccessNumericFilter

Filters for numeric or date values.

AccessOrderBy

Order bys define how rows will be sorted in the response. For example, ordering rows by descending access count is one ordering, and ordering rows by the country string is a different ordering.

AccessQuota

Current state of all quotas for this Analytics property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

AccessQuotaStatus

Current state for a particular quota group.

AccessRow

Access report data for each row.

AccessStringFilter

The filter for strings.

Account

A resource message representing a Google Analytics account.

AccountSummary

A virtual resource representing an overview of an account and all its child GA4 properties.

AcknowledgeUserDataCollectionRequest

Request message for AcknowledgeUserDataCollection RPC.

AcknowledgeUserDataCollectionResponse

Response message for AcknowledgeUserDataCollection RPC.

ArchiveCustomDimensionRequest

Request message for ArchiveCustomDimension RPC.

ArchiveCustomMetricRequest

Request message for ArchiveCustomMetric RPC.

ChangeHistoryChange

A description of a change to a single Google Analytics resource.

ChangeHistoryEvent

A set of changes within a Google Analytics account or its child properties that resulted from the same cause. Common causes would be updates made in the Google Analytics UI, changes from customer support, or automatic Google Analytics system changes.

ConversionEvent

A conversion event in a Google Analytics property.

CreateConversionEventRequest

Request message for CreateConversionEvent RPC

CreateCustomDimensionRequest

Request message for CreateCustomDimension RPC.

CreateCustomMetricRequest

Request message for CreateCustomMetric RPC.

CreateDataStreamRequest

Request message for CreateDataStream RPC.

CreateFirebaseLinkRequest

Request message for CreateFirebaseLink RPC

CreateGoogleAdsLinkRequest

Request message for CreateGoogleAdsLink RPC

CreateMeasurementProtocolSecretRequest

Request message for CreateMeasurementProtocolSecret RPC

CreatePropertyRequest

Request message for CreateProperty RPC.

CustomDimension

A definition for a CustomDimension.

CustomMetric

A definition for a custom metric.

DataRetentionSettings

Settings values for data retention. This is a singleton resource.

DataSharingSettings

A resource message representing data sharing settings of a Google Analytics account.

DataStream

A resource message representing a data stream.

DeleteAccountRequest

Request message for DeleteAccount RPC.

DeleteConversionEventRequest

Request message for DeleteConversionEvent RPC

DeleteDataStreamRequest

Request message for DeleteDataStream RPC.

DeleteFirebaseLinkRequest

Request message for DeleteFirebaseLink RPC

DeleteGoogleAdsLinkRequest

Request message for DeleteGoogleAdsLink RPC.

DeleteMeasurementProtocolSecretRequest

Request message for DeleteMeasurementProtocolSecret RPC

DeletePropertyRequest

Request message for DeleteProperty RPC.

FirebaseLink

A link between a GA4 property and a Firebase project.

GetAccountRequest

Request message for GetAccount RPC.

GetConversionEventRequest

Request message for GetConversionEvent RPC

GetCustomDimensionRequest

Request message for GetCustomDimension RPC.

GetCustomMetricRequest

Request message for GetCustomMetric RPC.

GetDataRetentionSettingsRequest

Request message for GetDataRetentionSettings RPC.

GetDataSharingSettingsRequest

Request message for GetDataSharingSettings RPC.

GetDataStreamRequest

Request message for GetDataStream RPC.

GetMeasurementProtocolSecretRequest

Request message for GetMeasurementProtocolSecret RPC.

GetPropertyRequest

Request message for GetProperty RPC.

GoogleAdsLink

A link between a GA4 property and a Google Ads account.

ListAccountSummariesRequest

Request message for ListAccountSummaries RPC.

ListAccountSummariesResponse

Response message for ListAccountSummaries RPC.

ListAccountsRequest

Request message for ListAccounts RPC.

ListAccountsResponse

Request message for ListAccounts RPC.

ListConversionEventsRequest

Request message for ListConversionEvents RPC

ListConversionEventsResponse

Response message for ListConversionEvents RPC.

ListCustomDimensionsRequest

Request message for ListCustomDimensions RPC.

ListCustomDimensionsResponse

Response message for ListCustomDimensions RPC.

ListCustomMetricsRequest

Request message for ListCustomMetrics RPC.

ListCustomMetricsResponse

Response message for ListCustomMetrics RPC.

ListDataStreamsRequest

Request message for ListDataStreams RPC.

ListDataStreamsResponse

Response message for ListDataStreams RPC.

ListFirebaseLinksRequest

Request message for ListFirebaseLinks RPC

ListFirebaseLinksResponse

Response message for ListFirebaseLinks RPC

ListGoogleAdsLinksRequest

Request message for ListGoogleAdsLinks RPC.

ListGoogleAdsLinksResponse

Response message for ListGoogleAdsLinks RPC.

ListMeasurementProtocolSecretsRequest

Request message for ListMeasurementProtocolSecret RPC

ListMeasurementProtocolSecretsResponse

Response message for ListMeasurementProtocolSecret RPC

ListPropertiesRequest

Request message for ListProperties RPC.

ListPropertiesResponse

Response message for ListProperties RPC.

MeasurementProtocolSecret

A secret value used for sending hits to Measurement Protocol.

NumericValue

To represent a number.

Property

A resource message representing a Google Analytics GA4 property.

PropertySummary

A virtual resource representing metadata for a GA4 property.

ProvisionAccountTicketRequest

Request message for ProvisionAccountTicket RPC.

ProvisionAccountTicketResponse

Response message for ProvisionAccountTicket RPC.

RunAccessReportRequest

The request for a Data Access Record Report.

RunAccessReportResponse

The customized Data Access Record Report response.

SearchChangeHistoryEventsRequest

Request message for SearchChangeHistoryEvents RPC.

SearchChangeHistoryEventsResponse

Response message for SearchAccounts RPC.

UpdateAccountRequest

Request message for UpdateAccount RPC.

UpdateConversionEventRequest

Request message for UpdateConversionEvent RPC

UpdateCustomDimensionRequest

Request message for UpdateCustomDimension RPC.

UpdateCustomMetricRequest

Request message for UpdateCustomMetric RPC.

UpdateDataRetentionSettingsRequest

Request message for UpdateDataRetentionSettings RPC.

UpdateDataStreamRequest

Request message for UpdateDataStream RPC.

UpdateGoogleAdsLinkRequest

Request message for UpdateGoogleAdsLink RPC

UpdateMeasurementProtocolSecretRequest

Request message for UpdateMeasurementProtocolSecret RPC

UpdatePropertyRequest

Request message for UpdateProperty RPC.

Enums

ActionType

Types of actions that may change a resource.

ActorType

Different kinds of actors that can make changes to Google Analytics resources.

ChangeHistoryResourceType

Types of resources whose changes may be returned from change history.

IndustryCategory

The category selected for this property, used for industry benchmarking.

PropertyType

Types of Property resources.

ServiceLevel

Various levels of service for Google Analytics.