Struct google_api_proto::google::analytics::admin::v1beta::analytics_admin_service_client::AnalyticsAdminServiceClient

pub struct AnalyticsAdminServiceClient<T> { /* private fields */ }

Service Interface for the Analytics Admin API (GA4).

Implementations

impl<T> AnalyticsAdminServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AnalyticsAdminServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_account( &mut self, request: impl IntoRequest<GetAccountRequest> ) -> Result<Response<Account>, Status>

Lookup for a single Account.

pub async fn list_accounts( &mut self, request: impl IntoRequest<ListAccountsRequest> ) -> Result<Response<ListAccountsResponse>, Status>

Returns all accounts accessible by the caller.

Note that these accounts might not currently have GA4 properties. Soft-deleted (ie: “trashed”) accounts are excluded by default. Returns an empty list if no relevant accounts are found.

pub async fn delete_account( &mut self, request: impl IntoRequest<DeleteAccountRequest> ) -> Result<Response<()>, Status>

Marks target Account as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI.

If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found.

pub async fn update_account( &mut self, request: impl IntoRequest<UpdateAccountRequest> ) -> Result<Response<Account>, Status>

Updates an account.

pub async fn provision_account_ticket( &mut self, request: impl IntoRequest<ProvisionAccountTicketRequest> ) -> Result<Response<ProvisionAccountTicketResponse>, Status>

Requests a ticket for creating an account.

pub async fn list_account_summaries( &mut self, request: impl IntoRequest<ListAccountSummariesRequest> ) -> Result<Response<ListAccountSummariesResponse>, Status>

Returns summaries of all accounts accessible by the caller.

pub async fn get_property( &mut self, request: impl IntoRequest<GetPropertyRequest> ) -> Result<Response<Property>, Status>

Lookup for a single “GA4” Property.

pub async fn list_properties( &mut self, request: impl IntoRequest<ListPropertiesRequest> ) -> Result<Response<ListPropertiesResponse>, Status>

Returns child Properties under the specified parent Account.

Only “GA4” properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: “trashed”) properties are excluded by default. Returns an empty list if no relevant properties are found.

pub async fn create_property( &mut self, request: impl IntoRequest<CreatePropertyRequest> ) -> Result<Response<Property>, Status>

Creates an “GA4” property with the specified location and attributes.

pub async fn delete_property( &mut self, request: impl IntoRequest<DeletePropertyRequest> ) -> Result<Response<Property>, Status>

Marks target Property as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI.

If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found, or is not a GA4 Property.

pub async fn update_property( &mut self, request: impl IntoRequest<UpdatePropertyRequest> ) -> Result<Response<Property>, Status>

Updates a property.

pub async fn create_firebase_link( &mut self, request: impl IntoRequest<CreateFirebaseLinkRequest> ) -> Result<Response<FirebaseLink>, Status>

Creates a FirebaseLink.

Properties can have at most one FirebaseLink.

pub async fn delete_firebase_link( &mut self, request: impl IntoRequest<DeleteFirebaseLinkRequest> ) -> Result<Response<()>, Status>

Deletes a FirebaseLink on a property

pub async fn list_firebase_links( &mut self, request: impl IntoRequest<ListFirebaseLinksRequest> ) -> Result<Response<ListFirebaseLinksResponse>, Status>

Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

pub async fn create_google_ads_link( &mut self, request: impl IntoRequest<CreateGoogleAdsLinkRequest> ) -> Result<Response<GoogleAdsLink>, Status>

Creates a GoogleAdsLink.

pub async fn update_google_ads_link( &mut self, request: impl IntoRequest<UpdateGoogleAdsLinkRequest> ) -> Result<Response<GoogleAdsLink>, Status>

Updates a GoogleAdsLink on a property

pub async fn delete_google_ads_link( &mut self, request: impl IntoRequest<DeleteGoogleAdsLinkRequest> ) -> Result<Response<()>, Status>

Deletes a GoogleAdsLink on a property

pub async fn list_google_ads_links( &mut self, request: impl IntoRequest<ListGoogleAdsLinksRequest> ) -> Result<Response<ListGoogleAdsLinksResponse>, Status>

Lists GoogleAdsLinks on a property.

pub async fn get_data_sharing_settings( &mut self, request: impl IntoRequest<GetDataSharingSettingsRequest> ) -> Result<Response<DataSharingSettings>, Status>

Get data sharing settings on an account. Data sharing settings are singletons.

pub async fn get_measurement_protocol_secret( &mut self, request: impl IntoRequest<GetMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Lookup for a single “GA4” MeasurementProtocolSecret.

pub async fn list_measurement_protocol_secrets( &mut self, request: impl IntoRequest<ListMeasurementProtocolSecretsRequest> ) -> Result<Response<ListMeasurementProtocolSecretsResponse>, Status>

Returns child MeasurementProtocolSecrets under the specified parent Property.

pub async fn create_measurement_protocol_secret( &mut self, request: impl IntoRequest<CreateMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Creates a measurement protocol secret.

pub async fn delete_measurement_protocol_secret( &mut self, request: impl IntoRequest<DeleteMeasurementProtocolSecretRequest> ) -> Result<Response<()>, Status>

Deletes target MeasurementProtocolSecret.

pub async fn update_measurement_protocol_secret( &mut self, request: impl IntoRequest<UpdateMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Updates a measurement protocol secret.

pub async fn acknowledge_user_data_collection( &mut self, request: impl IntoRequest<AcknowledgeUserDataCollectionRequest> ) -> Result<Response<AcknowledgeUserDataCollectionResponse>, Status>

Acknowledges the terms of user data collection for the specified property.

This acknowledgement must be completed (either in the Google Analytics UI or through this API) before MeasurementProtocolSecret resources may be created.

pub async fn search_change_history_events( &mut self, request: impl IntoRequest<SearchChangeHistoryEventsRequest> ) -> Result<Response<SearchChangeHistoryEventsResponse>, Status>

Searches through all changes to an account or its children given the specified set of filters.

pub async fn create_conversion_event( &mut self, request: impl IntoRequest<CreateConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Creates a conversion event with the specified attributes.

pub async fn update_conversion_event( &mut self, request: impl IntoRequest<UpdateConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Updates a conversion event with the specified attributes.

pub async fn get_conversion_event( &mut self, request: impl IntoRequest<GetConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Retrieve a single conversion event.

pub async fn delete_conversion_event( &mut self, request: impl IntoRequest<DeleteConversionEventRequest> ) -> Result<Response<()>, Status>

Deletes a conversion event in a property.

pub async fn list_conversion_events( &mut self, request: impl IntoRequest<ListConversionEventsRequest> ) -> Result<Response<ListConversionEventsResponse>, Status>

Returns a list of conversion events in the specified parent property.

Returns an empty list if no conversion events are found.

pub async fn create_custom_dimension( &mut self, request: impl IntoRequest<CreateCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Creates a CustomDimension.

pub async fn update_custom_dimension( &mut self, request: impl IntoRequest<UpdateCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Updates a CustomDimension on a property.

pub async fn list_custom_dimensions( &mut self, request: impl IntoRequest<ListCustomDimensionsRequest> ) -> Result<Response<ListCustomDimensionsResponse>, Status>

Lists CustomDimensions on a property.

pub async fn archive_custom_dimension( &mut self, request: impl IntoRequest<ArchiveCustomDimensionRequest> ) -> Result<Response<()>, Status>

Archives a CustomDimension on a property.

pub async fn get_custom_dimension( &mut self, request: impl IntoRequest<GetCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Lookup for a single CustomDimension.

pub async fn create_custom_metric( &mut self, request: impl IntoRequest<CreateCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Creates a CustomMetric.

pub async fn update_custom_metric( &mut self, request: impl IntoRequest<UpdateCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Updates a CustomMetric on a property.

pub async fn list_custom_metrics( &mut self, request: impl IntoRequest<ListCustomMetricsRequest> ) -> Result<Response<ListCustomMetricsResponse>, Status>

Lists CustomMetrics on a property.

pub async fn archive_custom_metric( &mut self, request: impl IntoRequest<ArchiveCustomMetricRequest> ) -> Result<Response<()>, Status>

Archives a CustomMetric on a property.

pub async fn get_custom_metric( &mut self, request: impl IntoRequest<GetCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Lookup for a single CustomMetric.

pub async fn get_data_retention_settings( &mut self, request: impl IntoRequest<GetDataRetentionSettingsRequest> ) -> Result<Response<DataRetentionSettings>, Status>

Returns the singleton data retention settings for this property.

pub async fn update_data_retention_settings( &mut self, request: impl IntoRequest<UpdateDataRetentionSettingsRequest> ) -> Result<Response<DataRetentionSettings>, Status>

Updates the singleton data retention settings for this property.

pub async fn create_data_stream( &mut self, request: impl IntoRequest<CreateDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Creates a DataStream.

pub async fn delete_data_stream( &mut self, request: impl IntoRequest<DeleteDataStreamRequest> ) -> Result<Response<()>, Status>

Deletes a DataStream on a property.

pub async fn update_data_stream( &mut self, request: impl IntoRequest<UpdateDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Updates a DataStream on a property.

pub async fn list_data_streams( &mut self, request: impl IntoRequest<ListDataStreamsRequest> ) -> Result<Response<ListDataStreamsResponse>, Status>

Lists DataStreams on a property.

pub async fn get_data_stream( &mut self, request: impl IntoRequest<GetDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Lookup for a single DataStream.

pub async fn run_access_report( &mut self, request: impl IntoRequest<RunAccessReportRequest> ) -> Result<Response<RunAccessReportResponse>, Status>

Returns a customized report of data access records. The report provides records of each time a user reads Google Analytics reporting data. Access records are retained for up to 2 years.

Data Access Reports can be requested for a property. The property must be in Google Analytics 360. This method is only available to Administrators.

These data access records include GA4 UI Reporting, GA4 UI Explorations, GA4 Data API, and other products like Firebase & Admob that can retrieve data from Google Analytics through a linkage. These records don’t include property configuration changes like adding a stream or changing a property’s time zone. For configuration change history, see searchChangeHistoryEvents.