Module google_api_proto::google::analytics::data::v1beta

Modules

audience_dimension_value

Nested message and enum types in AudienceDimensionValue.

audience_export

Nested message and enum types in AudienceExport.

beta_analytics_data_client

Generated client implementations.

cohorts_range

Nested message and enum types in CohortsRange.

dimension_expression

Nested message and enum types in DimensionExpression.

dimension_value

Nested message and enum types in DimensionValue.

filter

Nested message and enum types in Filter.

filter_expression

Nested message and enum types in FilterExpression.

metric_metadata

Nested message and enum types in MetricMetadata.

metric_value

Nested message and enum types in MetricValue.

numeric_value

Nested message and enum types in NumericValue.

order_by

Nested message and enum types in OrderBy.

response_meta_data

Nested message and enum types in ResponseMetaData.

Structs

AudienceDimension

An audience dimension is a user attribute. Specific user attributed are requested and then later returned in the QueryAudienceExportResponse.

AudienceDimensionValue

The value of a dimension.

AudienceExport

An audience export is a list of users in an audience at the time of the list’s creation. One audience may have multiple audience exports created for different days.

AudienceExportMetadata

This metadata is currently blank.

AudienceRow

Dimension value attributes for the audience user row.

BatchRunPivotReportsRequest

The batch request containing multiple pivot report requests.

BatchRunPivotReportsResponse

The batch response containing multiple pivot reports.

BatchRunReportsRequest

The batch request containing multiple report requests.

BatchRunReportsResponse

The batch response containing multiple reports.

CheckCompatibilityRequest

The request for compatibility information for a report’s dimensions and metrics. Check compatibility provides a preview of the compatibility of a report; fields shared with the runReport request should be the same values as in your runReport request.

CheckCompatibilityResponse

The compatibility response with the compatibility of each dimension & metric.

Cohort

Defines a cohort selection criteria. A cohort is a group of users who share a common characteristic. For example, users with the same firstSessionDate belong to the same cohort.

CohortReportSettings

Optional settings of a cohort report.

CohortSpec

The specification of cohorts for a cohort report.

CohortsRange

Configures the extended reporting date range for a cohort report. Specifies an offset duration to follow the cohorts over.

CreateAudienceExportRequest

A request to create a new audience export.

DateRange

A contiguous set of days: startDate, startDate + 1, …, endDate. Requests are allowed up to 4 date ranges.

Dimension

Dimensions are attributes of your data. For example, the dimension city indicates the city from which an event originates. Dimension values in report responses are strings; for example, the city could be “Paris” or “New York”. Requests are allowed up to 9 dimensions.

DimensionCompatibility

The compatibility for a single dimension.

DimensionExpression

Used to express a dimension which is the result of a formula of multiple dimensions. Example usages:

DimensionHeader

Describes a dimension column in the report. Dimensions requested in a report produce column entries within rows and DimensionHeaders. However, dimensions used exclusively within filters or expressions do not produce columns in a report; correspondingly, those dimensions do not produce headers.

DimensionMetadata

Explains a dimension.

DimensionValue

The value of a dimension.

Filter

An expression to filter dimension or metric values.

FilterExpression

To express dimension or metric filters. The fields in the same FilterExpression need to be either all dimensions or all metrics.

FilterExpressionList

A list of filter expressions.

GetAudienceExportRequest

A request to retrieve configuration metadata about a specific audience export.

GetMetadataRequest

Request for a property’s dimension and metric metadata.

ListAudienceExportsRequest

A request to list all audience exports for a property.

ListAudienceExportsResponse

A list of all audience exports for a property.

Metadata

The dimensions, metrics and comparisons currently accepted in reporting methods.

Metric

The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.

MetricCompatibility

The compatibility for a single metric.

MetricHeader

Describes a metric column in the report. Visible metrics requested in a report produce column entries within rows and MetricHeaders. However, metrics used exclusively within filters or expressions do not produce columns in a report; correspondingly, those metrics do not produce headers.

MetricMetadata

Explains a metric.

MetricValue

The value of a metric.

MinuteRange

A contiguous set of minutes: startMinutesAgo, startMinutesAgo + 1, …, endMinutesAgo. Requests are allowed up to 2 minute ranges.

NumericValue

To represent a number.

OrderBy

Order bys define how rows will be sorted in the response. For example, ordering rows by descending event count is one ordering, and ordering rows by the event name string is a different ordering.

Pivot

Describes the visible dimension columns and rows in the report response.

PivotDimensionHeader

Summarizes dimension values from a row for this pivot.

PivotHeader

Dimensions’ values in a single pivot.

PropertyQuota

Current state of all quotas for this Analytics Property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

QueryAudienceExportRequest

A request to list users in an audience export.

QueryAudienceExportResponse

A list of users in an audience export.

QuotaStatus

Current state for a particular quota group.

ResponseMetaData

Response’s metadata carrying additional information about the report content.

Row

Report data for each row. For example if RunReportRequest contains:

RunPivotReportRequest

The request to generate a pivot report.

RunPivotReportResponse

The response pivot report table corresponding to a pivot request.

RunRealtimeReportRequest

The request to generate a realtime report.

RunRealtimeReportResponse

The response realtime report table corresponding to a request.

RunReportRequest

The request to generate a report.

RunReportResponse

The response report table corresponding to a request.

SamplingMetadata

If this report results is sampled, this describes the percentage of events used in this report. Sampling is the practice of analyzing a subset of all data in order to uncover the meaningful information in the larger data set.

Enums

Compatibility

The compatibility types for a single dimension or metric.

MetricAggregation

Represents aggregation of metrics.

MetricType

A metric’s value type.

RestrictedMetricType

Categories of data that you may be restricted from viewing on certain GA4 properties.