Struct google_api_proto::google::analytics::data::v1beta::beta_analytics_data_client::BetaAnalyticsDataClient

pub struct BetaAnalyticsDataClient<T> { /* private fields */ }

Google Analytics reporting data service.

Implementations

impl<T> BetaAnalyticsDataClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> BetaAnalyticsDataClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn run_report( &mut self, request: impl IntoRequest<RunReportRequest> ) -> Result<Response<RunReportResponse>, Status>

Returns a customized report of your Google Analytics event data. Reports contain statistics derived from data collected by the Google Analytics tracking code. The data returned from the API is as a table with columns for the requested dimensions and metrics. Metrics are individual measurements of user activity on your property, such as active users or event count. Dimensions break down metrics across some common criteria, such as country or event name.

For a guide to constructing requests & understanding responses, see Creating a Report.

pub async fn run_pivot_report( &mut self, request: impl IntoRequest<RunPivotReportRequest> ) -> Result<Response<RunPivotReportResponse>, Status>

Returns a customized pivot report of your Google Analytics event data. Pivot reports are more advanced and expressive formats than regular reports. In a pivot report, dimensions are only visible if they are included in a pivot. Multiple pivots can be specified to further dissect your data.

pub async fn batch_run_reports( &mut self, request: impl IntoRequest<BatchRunReportsRequest> ) -> Result<Response<BatchRunReportsResponse>, Status>

Returns multiple reports in a batch. All reports must be for the same GA4 Property.

pub async fn batch_run_pivot_reports( &mut self, request: impl IntoRequest<BatchRunPivotReportsRequest> ) -> Result<Response<BatchRunPivotReportsResponse>, Status>

Returns multiple pivot reports in a batch. All reports must be for the same GA4 Property.

pub async fn get_metadata( &mut self, request: impl IntoRequest<GetMetadataRequest> ) -> Result<Response<Metadata>, Status>

Returns metadata for dimensions and metrics available in reporting methods. Used to explore the dimensions and metrics. In this method, a Google Analytics GA4 Property Identifier is specified in the request, and the metadata response includes Custom dimensions and metrics as well as Universal metadata.

For example if a custom metric with parameter name levels_unlocked is registered to a property, the Metadata response will contain customEvent:levels_unlocked. Universal metadata are dimensions and metrics applicable to any property such as country and totalUsers.

pub async fn run_realtime_report( &mut self, request: impl IntoRequest<RunRealtimeReportRequest> ) -> Result<Response<RunRealtimeReportResponse>, Status>

Returns a customized report of realtime event data for your property. Events appear in realtime reports seconds after they have been sent to the Google Analytics. Realtime reports show events and usage data for the periods of time ranging from the present moment to 30 minutes ago (up to 60 minutes for Google Analytics 360 properties).

For a guide to constructing realtime requests & understanding responses, see Creating a Realtime Report.

pub async fn check_compatibility( &mut self, request: impl IntoRequest<CheckCompatibilityRequest> ) -> Result<Response<CheckCompatibilityResponse>, Status>

This compatibility method lists dimensions and metrics that can be added to a report request and maintain compatibility. This method fails if the request’s dimensions and metrics are incompatible.

In Google Analytics, reports fail if they request incompatible dimensions and/or metrics; in that case, you will need to remove dimensions and/or metrics from the incompatible report until the report is compatible.

The Realtime and Core reports have different compatibility rules. This method checks compatibility for Core reports.

pub async fn create_audience_export( &mut self, request: impl IntoRequest<CreateAudienceExportRequest> ) -> Result<Response<Operation>, Status>

Creates an audience export for later retrieval. This method quickly returns the audience export’s resource name and initiates a long running asynchronous request to form an audience export. To export the users in an audience export, first create the audience export through this method and then send the audience resource name to the QueryAudienceExport method.

See Creating an Audience Export for an introduction to Audience Exports with examples.

An audience export is a snapshot of the users currently in the audience at the time of audience export creation. Creating audience exports for one audience on different days will return different results as users enter and exit the audience.

Audiences in Google Analytics 4 allow you to segment your users in the ways that are important to your business. To learn more, see https://support.google.com/analytics/answer/9267572. Audience exports contain the users in each audience.

Audience Export APIs have some methods at alpha and other methods at beta stability. The intention is to advance methods to beta stability after some feedback and adoption. To give your feedback on this API, complete the Google Analytics Audience Export API Feedback form.

pub async fn query_audience_export( &mut self, request: impl IntoRequest<QueryAudienceExportRequest> ) -> Result<Response<QueryAudienceExportResponse>, Status>

Retrieves an audience export of users. After creating an audience, the users are not immediately available for exporting. First, a request to CreateAudienceExport is necessary to create an audience export of users, and then second, this method is used to retrieve the users in the audience export.

See Creating an Audience Export for an introduction to Audience Exports with examples.

Audiences in Google Analytics 4 allow you to segment your users in the ways that are important to your business. To learn more, see https://support.google.com/analytics/answer/9267572.

Audience Export APIs have some methods at alpha and other methods at beta stability. The intention is to advance methods to beta stability after some feedback and adoption. To give your feedback on this API, complete the Google Analytics Audience Export API Feedback form.

pub async fn get_audience_export( &mut self, request: impl IntoRequest<GetAudienceExportRequest> ) -> Result<Response<AudienceExport>, Status>

Gets configuration metadata about a specific audience export. This method can be used to understand an audience export after it has been created.

See Creating an Audience Export for an introduction to Audience Exports with examples.

Audience Export APIs have some methods at alpha and other methods at beta stability. The intention is to advance methods to beta stability after some feedback and adoption. To give your feedback on this API, complete the Google Analytics Audience Export API Feedback form.

pub async fn list_audience_exports( &mut self, request: impl IntoRequest<ListAudienceExportsRequest> ) -> Result<Response<ListAudienceExportsResponse>, Status>

Lists all audience exports for a property. This method can be used for you to find and reuse existing audience exports rather than creating unnecessary new audience exports. The same audience can have multiple audience exports that represent the export of users that were in an audience on different days.

See Creating an Audience Export for an introduction to Audience Exports with examples.

Audience Export APIs have some methods at alpha and other methods at beta stability. The intention is to advance methods to beta stability after some feedback and adoption. To give your feedback on this API, complete the Google Analytics Audience Export API Feedback form.