Module google_api_proto::google::api::apikeys::v2

Modules

api_keys_client

Generated client implementations.

restrictions

Nested message and enum types in Restrictions.

Structs

AndroidApplication

Identifier of an Android application for key use.

AndroidKeyRestrictions

The Android apps that are allowed to use the key.

ApiTarget

A restriction for a specific service and optionally one or multiple specific methods. Both fields are case insensitive.

BrowserKeyRestrictions

The HTTP referrers (websites) that are allowed to use the key.

CreateKeyRequest

Request message for CreateKey method.

DeleteKeyRequest

Request message for DeleteKey method.

GetKeyRequest

Request message for GetKey method.

GetKeyStringRequest

Request message for GetKeyString method.

GetKeyStringResponse

Response message for GetKeyString method.

IosKeyRestrictions

The iOS apps that are allowed to use the key.

Key

The representation of a key managed by the API Keys API.

ListKeysRequest

Request message for ListKeys method.

ListKeysResponse

Response message for ListKeys method.

LookupKeyRequest

Request message for LookupKey method.

LookupKeyResponse

Response message for LookupKey method.

Restrictions

Describes the restrictions on the key.

ServerKeyRestrictions

The IP addresses of callers that are allowed to use the key.

UndeleteKeyRequest

Request message for UndeleteKey method.

UpdateKeyRequest

Request message for UpdateKey method.