Module google_api_proto::google::api::cloudquotas::v1

Modules

cloud_quotas_client

Generated client implementations.

quota_config

Nested message and enum types in QuotaConfig.

quota_increase_eligibility

Nested message and enum types in QuotaIncreaseEligibility.

quota_info

Nested message and enum types in QuotaInfo.

Structs

CreateQuotaPreferenceRequest

Message for creating a QuotaPreference

DimensionsInfo

The detailed quota information such as effective quota value for a combination of dimensions.

GetQuotaInfoRequest

Message for getting a QuotaInfo

GetQuotaPreferenceRequest

Message for getting a QuotaPreference

ListQuotaInfosRequest

Message for requesting list of QuotaInfos

ListQuotaInfosResponse

Message for response to listing QuotaInfos

ListQuotaPreferencesRequest

Message for requesting list of QuotaPreferences

ListQuotaPreferencesResponse

Message for response to listing QuotaPreferences

QuotaConfig

The preferred quota configuration.

QuotaDetails

The quota details for a map of dimensions.

QuotaIncreaseEligibility

Eligibility information regarding requesting increase adjustment of a quota.

QuotaInfo

QuotaInfo represents information about a particular quota for a given project, folder or organization.

QuotaPreference

QuotaPreference represents the preferred quota configuration specified for a project, folder or organization. There is only one QuotaPreference resource for a quota value targeting a unique set of dimensions.

RolloutInfo

[Output only] Rollout information of a quota.

UpdateQuotaPreferenceRequest

Message for updating a QuotaPreference

Enums

QuotaSafetyCheck

Enumerations of quota safety checks.