Module google_api_proto::google::api::expr::v1alpha1

Modules

constant

Nested message and enum types in Constant.

decl

Nested message and enum types in Decl.

eval_state

Nested message and enum types in EvalState.

explain

Nested message and enum types in Explain.

expr

Nested message and enum types in Expr.

expr_value

Nested message and enum types in ExprValue.

map_value

Nested message and enum types in MapValue.

source_info

Nested message and enum types in SourceInfo.

type

Nested message and enum types in Type.

value

Nested message and enum types in Value.

Structs

CheckedExpr

A CEL expression which has been successfully type checked.

Constant

Represents a primitive literal.

Decl

Represents a declaration of a named value or function.

EnumValue

An enum value.

ErrorSet

A set of errors.

EvalState

The state of an evaluation.

Explain

Values of intermediate expressions produced when evaluating expression. Deprecated, use EvalState instead.

Expr

An abstract representation of a common expression.

ExprValue

The value of an evaluated expression.

ListValue

A list.

MapValue

A map.

ParsedExpr

An expression together with source information as returned by the parser.

Reference

Describes a resolved reference to a declaration.

SourceInfo

Source information collected at parse time.

SourcePosition

A specific position in source.

Type

Represents a CEL type.

UnknownSet

A set of expressions for which the value is unknown.

Value

Represents a CEL value.