Module google_api_proto::google::api::servicecontrol::v1

Modules

check_error

Nested message and enum types in CheckError.

check_response

Nested message and enum types in CheckResponse.

distribution

Nested message and enum types in Distribution.

log_entry

Nested message and enum types in LogEntry.

metric_value

Nested message and enum types in MetricValue.

operation

Nested message and enum types in Operation.

quota_controller_client

Generated client implementations.

quota_error

Nested message and enum types in QuotaError.

quota_operation

Nested message and enum types in QuotaOperation.

report_response

Nested message and enum types in ReportResponse.

service_controller_client

Generated client implementations.

Structs

AllocateQuotaRequest

Request message for the AllocateQuota method.

AllocateQuotaResponse

Response message for the AllocateQuota method.

CheckError

Defines the errors to be returned in [google.api.servicecontrol.v1.CheckResponse.check_errors][google.api.servicecontrol.v1.CheckResponse.check_errors].

CheckRequest

Request message for the Check method.

CheckResponse

Response message for the Check method.

Distribution

Distribution represents a frequency distribution of double-valued sample points. It contains the size of the population of sample points plus additional optional information:

HttpRequest

A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.

LogEntry

An individual log entry.

LogEntryOperation

Additional information about a potentially long-running operation with which a log entry is associated.

LogEntrySourceLocation

Additional information about the source code location that produced the log entry.

MetricValue

Represents a single metric value.

MetricValueSet

Represents a set of metric values in the same metric. Each metric value in the set should have a unique combination of start time, end time, and label values.

Operation

Represents information regarding an operation.

QuotaError

Represents error information for [QuotaOperation][google.api.servicecontrol.v1.QuotaOperation].

QuotaOperation

Represents information regarding a quota operation.

ReportRequest

Request message for the Report method.

ReportResponse

Response message for the Report method.