Struct google_api_proto::google::api::servicecontrol::v1::HttpRequest

pub struct HttpRequest {

    pub request_method: String,
    pub request_url: String,
    pub request_size: i64,
    pub status: i32,
    pub response_size: i64,
    pub user_agent: String,
    pub remote_ip: String,
    pub server_ip: String,
    pub referer: String,
    pub latency: Option<Duration>,
    pub cache_lookup: bool,
    pub cache_hit: bool,
    pub cache_validated_with_origin_server: bool,
    pub cache_fill_bytes: i64,
    pub protocol: String,

}

A common proto for logging HTTP requests. Only contains semantics defined by the HTTP specification. Product-specific logging information MUST be defined in a separate message.

Fields
request_method: String

The request method. Examples: "GET", "HEAD", "PUT", "POST".

request_url: String

The scheme (http, https), the host name, the path, and the query portion of the URL that was requested. Example: "<http://example.com/some/info?color=red".>

request_size: i64

The size of the HTTP request message in bytes, including the request headers and the request body.

status: i32

The response code indicating the status of the response. Examples: 200, 404.

response_size: i64

The size of the HTTP response message sent back to the client, in bytes, including the response headers and the response body.

user_agent: String

The user agent sent by the client. Example: "Mozilla/4.0 (compatible; MSIE 6.0; Windows 98; Q312461; .NET CLR 1.0.3705)".

remote_ip: String

The IP address (IPv4 or IPv6) of the client that issued the HTTP request. Examples: "192.168.1.1", "FE80::0202:B3FF:FE1E:8329".

server_ip: String

The IP address (IPv4 or IPv6) of the origin server that the request was sent to.

referer: String

The referer URL of the request, as defined in HTTP/1.1 Header Field Definitions.

latency: Option<Duration>

The request processing latency on the server, from the time the request was received until the response was sent.

cache_lookup: bool

Whether or not a cache lookup was attempted.

cache_hit: bool

Whether or not an entity was served from cache (with or without validation).

cache_validated_with_origin_server: bool

Whether or not the response was validated with the origin server before being served from cache. This field is only meaningful if cache_hit is True.

cache_fill_bytes: i64

The number of HTTP response bytes inserted into cache. Set only when a cache fill was attempted.

protocol: String

Protocol used for the request. Examples: “HTTP/1.1”, “HTTP/2”, “websocket”