Module google_api_proto::google::api::servicemanagement::v1::rollout

Nested message and enum types in Rollout.

Structs

DeleteServiceStrategy

Strategy used to delete a service. This strategy is a placeholder only used by the system generated rollout to delete a service.

TrafficPercentStrategy

Strategy that specifies how clients of Google Service Controller want to send traffic to use different config versions. This is generally used by API proxy to split traffic based on your configured percentage for each config version.

Enums

RolloutStatus

Status of a Rollout.

Strategy

Strategy that defines which versions of service configurations should be pushed and how they should be used at runtime.