Struct google_api_proto::google::api::Service

pub struct Service {

    pub name: String,
    pub title: String,
    pub producer_project_id: String,
    pub id: String,
    pub apis: Vec<Api>,
    pub types: Vec<Type>,
    pub enums: Vec<Enum>,
    pub documentation: Option<Documentation>,
    pub backend: Option<Backend>,
    pub http: Option<Http>,
    pub quota: Option<Quota>,
    pub authentication: Option<Authentication>,
    pub context: Option<Context>,
    pub usage: Option<Usage>,
    pub endpoints: Vec<Endpoint>,
    pub control: Option<Control>,
    pub logs: Vec<LogDescriptor>,
    pub metrics: Vec<MetricDescriptor>,
    pub monitored_resources: Vec<MonitoredResourceDescriptor>,
    pub billing: Option<Billing>,
    pub logging: Option<Logging>,
    pub monitoring: Option<Monitoring>,
    pub system_parameters: Option<SystemParameters>,
    pub source_info: Option<SourceInfo>,
    pub publishing: Option<Publishing>,
    pub config_version: Option<u32>,

}

Service is the root object of Google API service configuration (service config). It describes the basic information about a logical service, such as the service name and the user-facing title, and delegates other aspects to sub-sections. Each sub-section is either a proto message or a repeated proto message that configures a specific aspect, such as auth. For more information, see each proto message definition.

Example:

 type: google.api.Service
 name: calendar.googleapis.com
 title: Google Calendar API
 apis:
 - name: google.calendar.v3.Calendar

 visibility:
   rules:
   - selector: "google.calendar.v3.*"
     restriction: PREVIEW
 backend:
   rules:
   - selector: "google.calendar.v3.*"
     address: calendar.example.com

 authentication:
   providers:
   - id: google_calendar_auth
     jwks_uri: <https://www.googleapis.com/oauth2/v1/certs>
     issuer: <https://securetoken.google.com>
   rules:
   - selector: "*"
     requirements:
       provider_id: google_calendar_auth

Fields
name: String

The service name, which is a DNS-like logical identifier for the service, such as calendar.googleapis.com. The service name typically goes through DNS verification to make sure the owner of the service also owns the DNS name.

title: String

The product title for this service, it is the name displayed in Google Cloud Console.

producer_project_id: String

The Google project that owns this service.

id: String

A unique ID for a specific instance of this message, typically assigned by the client for tracking purpose. Must be no longer than 63 characters and only lower case letters, digits, ‘.’, ‘_’ and ‘-’ are allowed. If empty, the server may choose to generate one instead.

apis: Vec<Api>

A list of API interfaces exported by this service. Only the name field of the [google.protobuf.Api][google.protobuf.Api] needs to be provided by the configuration author, as the remaining fields will be derived from the IDL during the normalization process. It is an error to specify an API interface here which cannot be resolved against the associated IDL files.

types: Vec<Type>

A list of all proto message types included in this API service. Types referenced directly or indirectly by the apis are automatically included. Messages which are not referenced but shall be included, such as types used by the google.protobuf.Any type, should be listed here by name by the configuration author. Example:

 types:
 - name: google.protobuf.Int32

enums: Vec<Enum>

A list of all enum types included in this API service. Enums referenced directly or indirectly by the apis are automatically included. Enums which are not referenced but shall be included should be listed here by name by the configuration author. Example:

 enums:
 - name: google.someapi.v1.SomeEnum

documentation: Option<Documentation>

Additional API documentation.

backend: Option<Backend>

API backend configuration.

http: Option<Http>

HTTP configuration.

quota: Option<Quota>

Quota configuration.

authentication: Option<Authentication>

Auth configuration.

context: Option<Context>

Context configuration.

usage: Option<Usage>

Configuration controlling usage of this service.

endpoints: Vec<Endpoint>

Configuration for network endpoints. If this is empty, then an endpoint with the same name as the service is automatically generated to service all defined APIs.

control: Option<Control>

Configuration for the service control plane.

logs: Vec<LogDescriptor>

Defines the logs used by this service.

metrics: Vec<MetricDescriptor>

Defines the metrics used by this service.

monitored_resources: Vec<MonitoredResourceDescriptor>

Defines the monitored resources used by this service. This is required by the [Service.monitoring][google.api.Service.monitoring] and [Service.logging][google.api.Service.logging] configurations.

billing: Option<Billing>

Billing configuration.

logging: Option<Logging>

Logging configuration.

monitoring: Option<Monitoring>

Monitoring configuration.

system_parameters: Option<SystemParameters>

System parameter configuration.

source_info: Option<SourceInfo>

Output only. The source information for this configuration if available.

publishing: Option<Publishing>

Settings for Google Cloud Client libraries generated from APIs defined as protocol buffers.

config_version: Option<u32>

Obsolete. Do not use.

This field has no semantic meaning. The service config compiler always sets this field to 3.