Struct google_api_proto::google::appengine::v1beta::authorized_certificates_client::AuthorizedCertificatesClient

pub struct AuthorizedCertificatesClient<T> { /* private fields */ }

Manages SSL certificates a user is authorized to administer. A user can administer any SSL certificates applicable to their authorized domains.

Implementations

impl<T> AuthorizedCertificatesClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> AuthorizedCertificatesClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_authorized_certificates( &mut self, request: impl IntoRequest<ListAuthorizedCertificatesRequest>, ) -> Result<Response<ListAuthorizedCertificatesResponse>, Status>

Lists all SSL certificates the user is authorized to administer.

pub async fn get_authorized_certificate( &mut self, request: impl IntoRequest<GetAuthorizedCertificateRequest>, ) -> Result<Response<AuthorizedCertificate>, Status>

Gets the specified SSL certificate.

pub async fn create_authorized_certificate( &mut self, request: impl IntoRequest<CreateAuthorizedCertificateRequest>, ) -> Result<Response<AuthorizedCertificate>, Status>

Uploads the specified SSL certificate.

pub async fn update_authorized_certificate( &mut self, request: impl IntoRequest<UpdateAuthorizedCertificateRequest>, ) -> Result<Response<AuthorizedCertificate>, Status>

Updates the specified SSL certificate. To renew a certificate and maintain its existing domain mappings, update certificate_data with a new certificate. The new certificate must be applicable to the same domains as the original certificate. The certificate display_name may also be updated.

pub async fn delete_authorized_certificate( &mut self, request: impl IntoRequest<DeleteAuthorizedCertificateRequest>, ) -> Result<Response<()>, Status>

Deletes the specified SSL certificate.