Struct google_api_proto::google::appengine::v1beta::Instance

pub struct Instance {

    pub name: String,
    pub id: String,
    pub app_engine_release: String,
    pub availability: i32,
    pub vm_name: String,
    pub vm_zone_name: String,
    pub vm_id: String,
    pub start_time: Option<Timestamp>,
    pub requests: i32,
    pub errors: i32,
    pub qps: f32,
    pub average_latency: i32,
    pub memory_usage: i64,
    pub vm_status: String,
    pub vm_debug_enabled: bool,
    pub vm_ip: String,
    pub vm_liveness: i32,

}

An Instance resource is the computing unit that App Engine uses to automatically scale an application.

Fields
name: String

Output only. Full path to the Instance resource in the API. Example: apps/myapp/services/default/versions/v1/instances/instance-1.

id: String

Output only. Relative name of the instance within the version. Example: instance-1.

app_engine_release: String

Output only. App Engine release this instance is running on.

availability: i32

Output only. Availability of the instance.

vm_name: String

Output only. Name of the virtual machine where this instance lives. Only applicable for instances in App Engine flexible environment.

vm_zone_name: String

Output only. Zone where the virtual machine is located. Only applicable for instances in App Engine flexible environment.

vm_id: String

Output only. Virtual machine ID of this instance. Only applicable for instances in App Engine flexible environment.

start_time: Option<Timestamp>

Output only. Time that this instance was started.

@OutputOnly

requests: i32

Output only. Number of requests since this instance was started.

errors: i32

Output only. Number of errors since this instance was started.

qps: f32

Output only. Average queries per second (QPS) over the last minute.

average_latency: i32

Output only. Average latency (ms) over the last minute.

memory_usage: i64

Output only. Total memory in use (bytes).

vm_status: String

Output only. Status of the virtual machine where this instance lives. Only applicable for instances in App Engine flexible environment.

vm_debug_enabled: bool

Output only. Whether this instance is in debug mode. Only applicable for instances in App Engine flexible environment.

vm_ip: String

Output only. The IP address of this instance. Only applicable for instances in App Engine flexible environment.

vm_liveness: i32

Output only. The liveness health check of this instance. Only applicable for instances in App Engine flexible environment.

Implementations

impl Instance

pub fn availability(&self) -> Availability

Returns the enum value of availability, or the default if the field is set to an invalid enum value.

pub fn set_availability(&mut self, value: Availability)

Sets availability to the provided enum value.

pub fn vm_liveness(&self) -> LivenessState

Returns the enum value of vm_liveness, or the default if the field is set to an invalid enum value.

pub fn set_vm_liveness(&mut self, value: LivenessState)

Sets vm_liveness to the provided enum value.