Module google_api_proto::google::apps::alertcenter::v1beta1

Modules

alert_center_service_client

Generated client implementations.

settings

Nested message and enum types in Settings.

Structs

Alert

An alert affecting a customer.

AlertFeedback

A customer feedback about an alert.

AlertMetadata

An alert metadata.

BatchDeleteAlertsRequest

A request to perform batch delete on alerts.

BatchDeleteAlertsResponse

Response to batch delete operation on alerts.

BatchUndeleteAlertsRequest

A request to perform batch undelete on alerts.

BatchUndeleteAlertsResponse

Response to batch undelete operation on alerts.

CreateAlertFeedbackRequest

A request to create a new alert feedback.

DeleteAlertRequest

A request to mark a specific alert for deletion.

GetAlertMetadataRequest

Get the alert metadata.

GetAlertRequest

Request for a specific alert.

GetSettingsRequest

Get the customer level settings.

ListAlertFeedbackRequest

An alert feedback listing request.

ListAlertFeedbackResponse

Response message for an alert feedback listing request.

ListAlertsRequest

An alert listing request.

ListAlertsResponse

Response message for an alert listing request.

Settings

Customer-level settings.

UndeleteAlertRequest

A request to undelete a specific alert that was marked for deletion.

UpdateSettingsRequest

Update the customer level settings.

Enums

AlertFeedbackType

The type of alert feedback.