Struct google_api_proto::google::apps::alertcenter::v1beta1::alert_center_service_client::AlertCenterServiceClient

pub struct AlertCenterServiceClient<T> { /* private fields */ }

Google Workspace Alert Center API (beta).

Implementations

impl<T> AlertCenterServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> AlertCenterServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_alerts(
&mut self,
request: impl IntoRequest<ListAlertsRequest>
) -> Result<Response<ListAlertsResponse>, Status>

Lists the alerts.

pub async fn get_alert(
&mut self,
request: impl IntoRequest<GetAlertRequest>
) -> Result<Response<Alert>, Status>

Gets the specified alert. Attempting to get a nonexistent alert returns NOT_FOUND error.

pub async fn delete_alert(
&mut self,
request: impl IntoRequest<DeleteAlertRequest>
) -> Result<Response<()>, Status>

Marks the specified alert for deletion. An alert that has been marked for deletion is removed from Alert Center after 30 days. Marking an alert for deletion has no effect on an alert which has already been marked for deletion. Attempting to mark a nonexistent alert for deletion results in a NOT_FOUND error.

pub async fn undelete_alert(
&mut self,
request: impl IntoRequest<UndeleteAlertRequest>
) -> Result<Response<Alert>, Status>

Restores, or “undeletes”, an alert that was marked for deletion within the past 30 days. Attempting to undelete an alert which was marked for deletion over 30 days ago (which has been removed from the Alert Center database) or a nonexistent alert returns a NOT_FOUND error. Attempting to undelete an alert which has not been marked for deletion has no effect.

pub async fn create_alert_feedback(
&mut self,
request: impl IntoRequest<CreateAlertFeedbackRequest>
) -> Result<Response<AlertFeedback>, Status>

Creates new feedback for an alert. Attempting to create a feedback for a non-existent alert returns NOT_FOUND error. Attempting to create a feedback for an alert that is marked for deletion returns `FAILED_PRECONDITION’ error.

pub async fn list_alert_feedback(
&mut self,
request: impl IntoRequest<ListAlertFeedbackRequest>
) -> Result<Response<ListAlertFeedbackResponse>, Status>

Lists all the feedback for an alert. Attempting to list feedbacks for a non-existent alert returns NOT_FOUND error.

pub async fn get_alert_metadata(
&mut self,
request: impl IntoRequest<GetAlertMetadataRequest>
) -> Result<Response<AlertMetadata>, Status>

Returns the metadata of an alert. Attempting to get metadata for a non-existent alert returns NOT_FOUND error.

pub async fn get_settings(
&mut self,
request: impl IntoRequest<GetSettingsRequest>
) -> Result<Response<Settings>, Status>

Returns customer-level settings.

pub async fn update_settings(
&mut self,
request: impl IntoRequest<UpdateSettingsRequest>
) -> Result<Response<Settings>, Status>

Updates the customer-level settings.

pub async fn batch_delete_alerts(
&mut self,
request: impl IntoRequest<BatchDeleteAlertsRequest>
) -> Result<Response<BatchDeleteAlertsResponse>, Status>

Performs batch delete operation on alerts.

pub async fn batch_undelete_alerts(
&mut self,
request: impl IntoRequest<BatchUndeleteAlertsRequest>
) -> Result<Response<BatchUndeleteAlertsResponse>, Status>

Performs batch undelete operation on alerts.