Module google_api_proto::google::apps::card::v1::card

Nested message and enum types in Card.

Structs

CardAction

A card action is the action associated with the card. For example, an invoice card might include actions such as delete invoice, email invoice, or open the invoice in a browser.

CardFixedFooter

A persistent (sticky) footer that that appears at the bottom of the card.

CardHeader

Represents a card header. For an example in Google Chat apps, see Card header.

Section

A section contains a collection of widgets that are rendered vertically in the order that they’re specified.

Enums

DisplayStyle

In Google Workspace Add-ons, determines how a card is displayed.

DividerStyle

The divider style of a card. Currently only used for dividers betweens card sections.