Module google_api_proto::google::apps::drive::labels::v2

Modules

field

Nested message and enum types in Field.

invalid_argument

Nested message and enum types in InvalidArgument.

label

Nested message and enum types in Label.

label_permission

Nested message and enum types in LabelPermission.

label_service_client

Generated client implementations.

lifecycle

Nested message and enum types in Lifecycle.

list_labels_request

Nested message and enum types in ListLabelsRequest.

precondition_failure

Nested message and enum types in PreconditionFailure.

Structs

BadgeColors

The color derived from BadgeConfig and changed to the closest recommended supported color.

BadgeConfig

Badge status of the label.

ExceptionDetail

Normalized internal-only message that identifies the exact exception that caused the error on the server.

Field

Defines a field that has a display name, data type, and other configuration options. This field defines the kind of metadata that may be set on a Drive item.

GetLabelRequest

Request to get a label by resource name.

InvalidArgument

Describes violations in a request to create or update a Label or its Fields.

Label

A label defines a taxonomy that can be applied to Drive items in order to organize and search across items. Labels can be simple strings, or can contain fields that describe additional metadata that can be further used to organize and search Drive items.

LabelPermission

The permission that applies to a principal (user, group, audience) on a label.

Lifecycle

The lifecycle state of an object, such as label, field, or choice. The lifecycle enforces the following transitions:

ListLabelsRequest

Request to list labels available to the current user.

ListLabelsResponse

Response for listing Labels.

LockStatus

Contains information about whether a label component should be considered locked.

PreconditionFailure

Describes what preconditions have failed.

UserInfo

Information about a user.

Enums

ExceptionType

Every ExceptionType maps to one and only one Exception class. This allows internal clients to identify the exact server exception that caused the error for debugging and logging purposes. Add new ExceptionTypes to EXCEPTION_TYPE_TO_ERROR_CODE_MAP in j/c/g/apps/boq/metadata/model/service/exceptions/CategoryExceptionHelper

LabelView

Resource view that can be applied to label responses. The default value LABEL_VIEW_BASIC implies the field mask: name,id,revision_id,label_type,properties.*\