Module google_api_proto::google::apps::drive::labels::v2::field

Nested message and enum types in Field.

Modules

date_options

Nested message and enum types in DateOptions.

selection_options

Nested message and enum types in SelectionOptions.

Structs

AppliedCapabilities

The capabilities related to this field on applied metadata.

DateOptions

Options for the date field type.

DisplayHints

UI display hints for rendering a field.

IntegerOptions

Options for the Integer field type.

ListOptions

Options for a multi-valued variant of an associated field type.

Properties

The basic properties of the field.

SchemaCapabilities

The capabilities related to this field when editing the field.

SelectionOptions

Options for the selection field type.

TextOptions

Options for the Text field type.

UserOptions

Options for the user field type.

Enums

Type

The data type and options of this field. Once published, the data type cannot be changed.