Struct google_api_proto::google::apps::drive::labels::v2::label_service_client::LabelServiceClient

pub struct LabelServiceClient<T> { /* private fields */ }

Manage metadata taxonomies based on Labels and Fields that may be used within Google Drive to organize and find files using custom metadata.

Implementations

impl<T> LabelServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> LabelServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_user_capabilities( &mut self, request: impl IntoRequest<GetUserCapabilitiesRequest>, ) -> Result<Response<UserCapabilities>, Status>

Gets the user capabilities.

pub async fn list_labels( &mut self, request: impl IntoRequest<ListLabelsRequest>, ) -> Result<Response<ListLabelsResponse>, Status>

List labels.

pub async fn get_label( &mut self, request: impl IntoRequest<GetLabelRequest>, ) -> Result<Response<Label>, Status>

Get a label by its resource name. Resource name may be any of:

labels/{id} - See labels/{id}@latest

labels/{id}@latest - Gets the latest revision of the label.

labels/{id}@published - Gets the current published revision of the label.

labels/{id}@{revision_id} - Gets the label at the specified revision ID.

pub async fn get_label_limits( &mut self, request: impl IntoRequest<GetLabelLimitsRequest>, ) -> Result<Response<LabelLimits>, Status>

Get the constraints on the structure of a Label; such as, the maximum number of Fields allowed and maximum length of the label title.

pub async fn create_label( &mut self, request: impl IntoRequest<CreateLabelRequest>, ) -> Result<Response<Label>, Status>

Creates a new Label.

pub async fn delta_update_label( &mut self, request: impl IntoRequest<DeltaUpdateLabelRequest>, ) -> Result<Response<DeltaUpdateLabelResponse>, Status>

Updates a single Label by applying a set of update requests resulting in a new draft revision. The batch update is all-or-nothing: If any of the update requests are invalid, no changes are applied. The resulting draft revision must be published before the changes may be used with Drive Items.

pub async fn update_label_copy_mode( &mut self, request: impl IntoRequest<UpdateLabelCopyModeRequest>, ) -> Result<Response<Label>, Status>

Updates a Label’s CopyMode. Changes to this policy are not revisioned, do not require publishing, and take effect immediately.

pub async fn publish_label( &mut self, request: impl IntoRequest<PublishLabelRequest>, ) -> Result<Response<Label>, Status>

Publish all draft changes to the Label. Once published, the Label may not return to its draft state. See google.apps.drive.labels.v2.Lifecycle for more information.

Publishing a Label will result in a new published revision. All previous draft revisions will be deleted. Previous published revisions will be kept but are subject to automated deletion as needed.

Once published, some changes are no longer permitted. Generally, any change that would invalidate or cause new restrictions on existing metadata related to the Label will be rejected. For example, the following changes to a Label will be rejected after the Label is published:

The label cannot be directly deleted. It must be disabled first, then deleted.

Field.FieldType cannot be changed.

Changes to Field validation options cannot reject something that was previously accepted.

Reducing the max entries.

pub async fn disable_label( &mut self, request: impl IntoRequest<DisableLabelRequest>, ) -> Result<Response<Label>, Status>

Disable a published Label. Disabling a Label will result in a new disabled published revision based on the current published revision. If there is a draft revision, a new disabled draft revision will be created based on the latest draft revision. Older draft revisions will be deleted.

Once disabled, a label may be deleted with DeleteLabel.

pub async fn enable_label( &mut self, request: impl IntoRequest<EnableLabelRequest>, ) -> Result<Response<Label>, Status>

Enable a disabled Label and restore it to its published state. This will result in a new published revision based on the current disabled published revision. If there is an existing disabled draft revision, a new revision will be created based on that draft and will be enabled.

pub async fn delete_label( &mut self, request: impl IntoRequest<DeleteLabelRequest>, ) -> Result<Response<()>, Status>

Permanently deletes a Label and related metadata on Drive Items.

Once deleted, the Label and related Drive item metadata will be deleted. Only draft Labels, and disabled Labels may be deleted.

pub async fn list_label_permissions( &mut self, request: impl IntoRequest<ListLabelPermissionsRequest>, ) -> Result<Response<ListLabelPermissionsResponse>, Status>

Lists a Label’s permissions.

pub async fn create_label_permission( &mut self, request: impl IntoRequest<CreateLabelPermissionRequest>, ) -> Result<Response<LabelPermission>, Status>

Updates a Label’s permissions. If a permission for the indicated principal doesn’t exist, a new Label Permission is created, otherwise the existing permission is updated. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

pub async fn update_label_permission( &mut self, request: impl IntoRequest<UpdateLabelPermissionRequest>, ) -> Result<Response<LabelPermission>, Status>

Updates a Label’s permissions. If a permission for the indicated principal doesn’t exist, a new Label Permission is created, otherwise the existing permission is updated. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

pub async fn delete_label_permission( &mut self, request: impl IntoRequest<DeleteLabelPermissionRequest>, ) -> Result<Response<()>, Status>

Deletes a Label’s permission. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

pub async fn batch_update_label_permissions( &mut self, request: impl IntoRequest<BatchUpdateLabelPermissionsRequest>, ) -> Result<Response<BatchUpdateLabelPermissionsResponse>, Status>

Updates Label permissions. If a permission for the indicated principal doesn’t exist, a new Label Permission is created, otherwise the existing permission is updated. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

pub async fn batch_delete_label_permissions( &mut self, request: impl IntoRequest<BatchDeleteLabelPermissionsRequest>, ) -> Result<Response<()>, Status>

Deletes Label permissions. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

pub async fn list_label_locks( &mut self, request: impl IntoRequest<ListLabelLocksRequest>, ) -> Result<Response<ListLabelLocksResponse>, Status>

Lists the LabelLocks on a Label.