Module google_api_proto::google::apps::drive::labels::v2beta

Modules

delta_update_label_request

Nested message and enum types in DeltaUpdateLabelRequest.

delta_update_label_response

Nested message and enum types in DeltaUpdateLabelResponse.

field

Nested message and enum types in Field.

invalid_argument

Nested message and enum types in InvalidArgument.

label

Nested message and enum types in Label.

label_lock

Nested message and enum types in LabelLock.

label_permission

Nested message and enum types in LabelPermission.

label_service_client

Generated client implementations.

lifecycle

Nested message and enum types in Lifecycle.

list_labels_request

Nested message and enum types in ListLabelsRequest.

precondition_failure

Nested message and enum types in PreconditionFailure.

write_control

Nested message and enum types in WriteControl.

Structs

BadgeColors

The color derived from BadgeConfig and changed to the closest recommended supported color.

BadgeConfig

Badge status of the label.

BatchDeleteLabelPermissionsRequest

Deletes one of more Label Permissions.

BatchUpdateLabelPermissionsRequest

Updates one or more Label Permissions.

BatchUpdateLabelPermissionsResponse

Response for updating one or more Label Permissions.

CreateLabelPermissionRequest

Creates or updates a permission on the Label. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

CreateLabelRequest

Request to create a Label.

DateLimits

Limits for date Field type.

DeleteLabelPermissionRequest

Deletes a Label Permission. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

DeleteLabelRequest

Request to delete a label.

DeltaUpdateLabelRequest

The set of requests for updating aspects of a Label. If any request is not valid, no requests will be applied.

DeltaUpdateLabelResponse

Response for Label update.

DisableLabelRequest

Request to deprecate a published Label.

EnableLabelRequest

Request to enable a label.

ExceptionDetail

Exception detail.

Field

Defines a field that has a display name, data type, and other configuration options. This field defines the kind of metadata that may be set on a Drive item.

FieldLimits

Field constants governing the structure of a Field; such as, the maximum title length, minimum and maximum field values or length, etc.

GetLabelLimitsRequest

Request to get the limits for a Label.

GetLabelRequest

Request to get a label by resource name.

GetUserCapabilitiesRequest

Request to get the capabilities for a user.

IntegerLimits

Limits for integer Field type.

InvalidArgument

Describes violations in a request to create or update a Label or its Fields.

Label

A label defines a taxonomy that can be applied to Drive items in order to organize and search across items. Labels can be simple strings, or can contain fields that describe additional metadata that can be further used to organize and search Drive items.

LabelLimits

Label constraints governing the structure of a Label; such as, the maximum number of Fields allowed and maximum length of the label title.

LabelLock

A Lock that can be applied to a Label, Field, or Choice.

LabelPermission

The permission that applies to a principal (user, group, audience) on a label.

Lifecycle

The lifecycle state of an object, such as label, field, or choice. The lifecycle enforces the following transitions:

ListLabelLocksRequest

A request to list the LabelLocks on a Label.

ListLabelLocksResponse

The response to a ListLabelLocksRequest.

ListLabelPermissionsRequest

Request to list the permissions on a Label.

ListLabelPermissionsResponse

Response for listing the permissions on a Label.

ListLabelsRequest

Request to list labels available to the current user.

ListLabelsResponse

Response for listing Labels.

ListLimits

Limits for list-variant of a Field type.

LockStatus

Contains information about whether a label component should be considered locked.

LongTextLimits

Limits for long text Field type.

PreconditionFailure

Describes what preconditions have failed.

PublishLabelRequest

Request to publish a label.

SelectionLimits

Limits for selection Field type.

TextLimits

Limits for text Field type.

UpdateLabelCopyModeRequest

Request to update the CopyMode of the given Label. Changes to this policy are not revisioned, do not require publishing, and take effect immediately. \

UpdateLabelPermissionRequest

Updates a Label Permission. Permissions affect the Label resource as a whole, are not revisioned, and do not require publishing.

UserCapabilities

The capabilities of a user.

UserInfo

Information about a user.

UserLimits

Limits for Field.Type.USER.

WriteControl

Provides control over how write requests are executed. When not specified, the last write wins.

Enums

ExceptionType

Every ExceptionType maps to one and only one Exception class. This allows internal clients to identify the exact server exception that caused the error for debugging and logging purposes. Add new ExceptionTypes to EXCEPTION_TYPE_TO_ERROR_CODE_MAP in j/c/g/apps/boq/metadata/model/service/exceptions/CategoryExceptionHelper

LabelView

Resource view that can be applied to label responses. The default value LABEL_VIEW_BASIC implies the field mask: name,id,revision_id,label_type,properties.*\