Module google_api_proto::google::apps::events::subscriptions::v1

Modules

notification_endpoint

Nested message and enum types in NotificationEndpoint.

subscription

Nested message and enum types in Subscription.

subscriptions_service_client

Generated client implementations.

Structs

CreateSubscriptionMetadata

Metadata for CreateSubscription LRO.

CreateSubscriptionRequest

The request message for [SubscriptionsService.CreateSubscription][google.apps.events.subscriptions.v1.SubscriptionsService.CreateSubscription].

DeleteSubscriptionMetadata

Metadata for DeleteSubscription LRO.

DeleteSubscriptionRequest

The request message for [SubscriptionsService.DeleteSubscription][google.apps.events.subscriptions.v1.SubscriptionsService.DeleteSubscription].

GetSubscriptionRequest

The request message for [SubscriptionsService.GetSubscription][google.apps.events.subscriptions.v1.SubscriptionsService.GetSubscription].

ListSubscriptionsRequest

The request message for [SubscriptionsService.ListSubscriptions][google.apps.events.subscriptions.v1.SubscriptionsService.ListSubscriptions].

ListSubscriptionsResponse

The response message for [SubscriptionsService.ListSubscriptions][google.apps.events.subscriptions.v1.SubscriptionsService.ListSubscriptions].

NotificationEndpoint

The endpoint where the subscription delivers events.

PayloadOptions

Options about what data to include in the event payload. Only supported for Google Chat events.

ReactivateSubscriptionMetadata

Metadata for ReactivateSubscription LRO.

ReactivateSubscriptionRequest

The request message for [SubscriptionsService.ReactivateSubscription][google.apps.events.subscriptions.v1.SubscriptionsService.ReactivateSubscription].

Subscription

A subscription to receive events about a Google Workspace resource. To learn more about subscriptions, see the Google Workspace Events API overview.

UpdateSubscriptionMetadata

Metadata for UpdateSubscription LRO.

UpdateSubscriptionRequest

The request message for [SubscriptionsService.UpdateSubscription][google.apps.events.subscriptions.v1.SubscriptionsService.UpdateSubscription].