Module google_api_proto::google::apps::meet::v2

Modules

conference_records_service_client

Generated client implementations.

participant

Nested message and enum types in Participant.

recording

Nested message and enum types in Recording.

space_config

Nested message and enum types in SpaceConfig.

spaces_service_client

Generated client implementations.

transcript

Nested message and enum types in Transcript.

Structs

ActiveConference

Active conference.

AnonymousUser

User who joins anonymously (meaning not signed into a Google Account).

ConferenceRecord

Single instance of a meeting held in a space.

CreateSpaceRequest

Request to create a space.

DocsDestination

Google Docs location where the transcript file is saved.

DriveDestination

Export location where a recording file is saved in Google Drive.

EndActiveConferenceRequest

Request to end an ongoing conference of a space.

GetConferenceRecordRequest

Request to get a conference record.

GetParticipantRequest

Request to get a participant.

GetParticipantSessionRequest

Request to get a participant session.

GetRecordingRequest

Request message for GetRecording method.

GetSpaceRequest

Request to get a space.

GetTranscriptEntryRequest

Request for GetTranscriptEntry method.

GetTranscriptRequest

Request for GetTranscript method.

ListConferenceRecordsRequest

Request to fetch list of conference records per user.

ListConferenceRecordsResponse

Response of ListConferenceRecords method.

ListParticipantSessionsRequest

Request to fetch list of participant sessions per conference record, per participant.

ListParticipantSessionsResponse

Response of ListParticipants method.

ListParticipantsRequest

Request to fetch list of participants per conference.

ListParticipantsResponse

Response of ListParticipants method.

ListRecordingsRequest

Request for ListRecordings method.

ListRecordingsResponse

Response for ListRecordings method.

ListTranscriptEntriesRequest

Request for ListTranscriptEntries method.

ListTranscriptEntriesResponse

Response for ListTranscriptEntries method.

ListTranscriptsRequest

Request for ListTranscripts method.

ListTranscriptsResponse

Response for ListTranscripts method.

Participant

User who attended or is attending a conference.

ParticipantSession

Refers to each unique join or leave session when a user joins a conference from a device. Note that any time a user joins the conference a new unique ID is assigned. That means if a user joins a space multiple times from the same device, they’re assigned different IDs, and are also be treated as different participant sessions.

PhoneUser

User dialing in from a phone where the user’s identity is unknown because they haven’t signed in with a Google Account.

Recording

Metadata about a recording created during a conference.

SignedinUser

A signed-in user can be: a) An individual joining from a personal computer, mobile device, or through companion mode. b) A robot account used by conference room devices.

Space

Virtual place where conferences are held. Only one active conference can be held in one space at any given time.

SpaceConfig

The configuration pertaining to a meeting space.

Transcript

Metadata for a transcript generated from a conference. It refers to the ASR (Automatic Speech Recognition) result of user’s speech during the conference.

TranscriptEntry

Single entry for one user’s speech during a transcript session.

UpdateSpaceRequest

Request to update a space.