Module google_api_proto::google::bigtable::v2::read_change_stream_response

Nested message and enum types in ReadChangeStreamResponse.

Modules

data_change

Nested message and enum types in DataChange.

mutation_chunk

Nested message and enum types in MutationChunk.

Structs

CloseStream

A message indicating that the client should stop reading from the stream. If status is OK and continuation_tokens & new_partitions are empty, the stream has finished (for example if there was an end_time specified). If continuation_tokens & new_partitions are present, then a change in partitioning requires the client to open a new stream for each token to resume reading. Example: [B, D) ends | v new_partitions: [A, C) [C, E) continuation_tokens.partitions: [B,C) [C,D) ^—^ ^—^ ^ ^ | | | StreamContinuationToken 2 | StreamContinuationToken 1 To read the new partition [A,C), supply the continuation tokens whose ranges cover the new partition, for example ContinuationToken[A,B) & ContinuationToken[B,C).

DataChange

A message corresponding to one or more mutations to the partition being streamed. A single logical DataChange message may also be split across a sequence of multiple individual messages. Messages other than the first in a sequence will only have the type and chunks fields populated, with the final message in the sequence also containing done set to true.

Heartbeat

A periodic message with information that can be used to checkpoint the state of a stream.

MutationChunk

A partial or complete mutation.

Enums

StreamRecord

The data or control message on the stream.