Module google_api_proto::google::bigtable::v2::type

Nested message and enum types in Type.

Modules

aggregate

Nested message and enum types in Aggregate.

bytes

Nested message and enum types in Bytes.

int64

Nested message and enum types in Int64.

string

Nested message and enum types in String.

struct

Nested message and enum types in Struct.

Structs

Aggregate

A value that combines incremental updates into a summarized value.

Array

An ordered list of elements of a given type. Values of type Array are stored in Value.array_value.

Bool

bool Values of type Bool are stored in Value.bool_value.

Bytes

Bytes Values of type Bytes are stored in Value.bytes_value.

Date

Date Values of type Date are stored in Value.date_value.

Float32

Float32 Values of type Float32 are stored in Value.float_value.

Float64

Float64 Values of type Float64 are stored in Value.float_value.

Int64

Int64 Values of type Int64 are stored in Value.int_value.

Map

A mapping of keys to values of a given type. Values of type Map are stored in a Value.array_value where each entry is another Value.array_value with two elements (the key and the value, in that order). Normally encoded Map values won’t have repeated keys, however, clients are expected to handle the case in which they do. If the same key appears multiple times, the last value takes precedence.

String

String Values of type String are stored in Value.string_value.

Struct

A structured data value, consisting of fields which map to dynamically typed values. Values of type Struct are stored in Value.array_value where entries are in the same order and number as field_types.

Timestamp

Timestamp Values of type Timestamp are stored in Value.timestamp_value.

Enums

Kind

The kind of type that this represents.