Module google_api_proto::google::chat::v1

Modules

accessory_widget

Nested message and enum types in AccessoryWidget.

action_response

Nested message and enum types in ActionResponse.

annotation

Nested message and enum types in Annotation.

attachment

Nested message and enum types in Attachment.

chat_service_client

Generated client implementations.

contextual_add_on_markup

Nested message and enum types in ContextualAddOnMarkup.

create_message_request

Nested message and enum types in CreateMessageRequest.

deletion_metadata

Nested message and enum types in DeletionMetadata.

dialog_action

Nested message and enum types in DialogAction.

emoji

Nested message and enum types in Emoji.

membership

Nested message and enum types in Membership.

rich_link_metadata

Nested message and enum types in RichLinkMetadata.

slash_command_metadata

Nested message and enum types in SlashCommandMetadata.

space

Nested message and enum types in Space.

space_event

Nested message and enum types in SpaceEvent.

user

Nested message and enum types in User.

user_mention_metadata

Nested message and enum types in UserMentionMetadata.

widget_markup

Nested message and enum types in WidgetMarkup.

Structs

AccessoryWidget

One or more interactive widgets that appear at the bottom of a message. For details, see Add interactive widgets at the bottom of a message.

ActionResponse

Parameters that a Chat app can use to configure how its response is posted.

ActionStatus

Represents the status for a request to either invoke or submit a dialog.

Annotation

Output only. Annotations associated with the plain-text body of the message. To add basic formatting to a text message, see Format text messages.

AttachedGif

A GIF image that’s specified by a URL.

Attachment

An attachment in Google Chat.

AttachmentDataRef

A reference to the attachment data.

CardWithId

A card in a Google Chat message.

CompleteImportSpaceRequest

Request message for completing the import process for a space.

CompleteImportSpaceResponse

Response message for completing the import process for a space.

ContextualAddOnMarkup

The markup for developers to specify the contents of a contextual AddOn.

CreateMembershipRequest

Request message for creating a membership.

CreateMessageRequest

Creates a message.

CreateReactionRequest

Creates a reaction to a message.

CreateSpaceRequest

A request to create a named space.

CustomEmoji

Represents a custom emoji.

DeleteMembershipRequest

Request to delete a membership in a space.

DeleteMessageRequest

Request to delete a message.

DeleteReactionRequest

Deletes a reaction to a message.

DeleteSpaceRequest

Request for deleting a space.

DeletionMetadata

Information about a deleted message. A message is deleted when delete_time is set.

Dialog

Wrapper around the card body of the dialog.

DialogAction

Contains a dialog and request status code.

DriveDataRef

A reference to the data of a drive attachment.

DriveLinkData

Data for Google Drive links.

Emoji

An emoji that is used as a reaction to a message.

EmojiReactionSummary

The number of people who reacted to a message with a specific emoji.

FindDirectMessageRequest

A request to get direct message space based on the user resource.

GetAttachmentRequest

Request to get an attachment.

GetMembershipRequest

Request to get a membership of a space.

GetMessageRequest

Request to get a message.

GetSpaceEventRequest

Request message for getting a space event.

GetSpaceReadStateRequest

Request message for GetSpaceReadState API.

GetSpaceRequest

A request to return a single space.

GetThreadReadStateRequest

Request message for GetThreadReadStateRequest API.

Group

A Google Group in Google Chat.

ListMembershipsRequest

Request message for listing memberships.

ListMembershipsResponse

Response to list memberships of the space.

ListMessagesRequest

Lists messages in the specified space, that the user is a member of.

ListMessagesResponse

Response message for listing messages.

ListReactionsRequest

Lists reactions to a message.

ListReactionsResponse

Response to a list reactions request.

ListSpaceEventsRequest

Request message for listing space events.

ListSpaceEventsResponse

Response message for listing space events.

ListSpacesRequest

A request to list the spaces the caller is a member of.

ListSpacesResponse

The response for a list spaces request.

MatchedUrl

A matched URL in a Chat message. Chat apps can preview matched URLs. For more information, see Preview links.

Membership

Represents a membership relation in Google Chat, such as whether a user or Chat app is invited to, part of, or absent from a space.

MembershipBatchCreatedEventData

Event payload for multiple new memberships.

MembershipBatchDeletedEventData

Event payload for multiple deleted memberships.

MembershipBatchUpdatedEventData

Event payload for multiple updated memberships.

MembershipCreatedEventData

Event payload for a new membership.

MembershipDeletedEventData

Event payload for a deleted membership.

MembershipUpdatedEventData

Event payload for an updated membership.

Message

A message in a Google Chat space.

MessageBatchCreatedEventData

Event payload for multiple new messages.

MessageBatchDeletedEventData

Event payload for multiple deleted messages.

MessageBatchUpdatedEventData

Event payload for multiple updated messages.

MessageCreatedEventData

Event payload for a new message.

MessageDeletedEventData

Event payload for a deleted message.

MessageUpdatedEventData

Event payload for an updated message.

QuotedMessageMetadata

Information about a quoted message.

Reaction

A reaction to a message.

ReactionBatchCreatedEventData

Event payload for multiple new reactions.

ReactionBatchDeletedEventData

Event payload for multiple deleted reactions.

ReactionCreatedEventData

Event payload for a new reaction.

ReactionDeletedEventData

Event payload for a deleted reaction.

RichLinkMetadata

A rich link to a resource.

SetUpSpaceRequest

Request to create a space and add specified users to it.

SlashCommand

A slash command in Google Chat.

SlashCommandMetadata

Annotation metadata for slash commands (/).

Space

A space in Google Chat. Spaces are conversations between two or more users or 1:1 messages between a user and a Chat app.

SpaceBatchUpdatedEventData

Event payload for multiple updates to a space.

SpaceEvent

An event that represents a change or activity in a Google Chat space. To learn more, see Work with events from Google Chat.

SpaceReadState

A user’s read state within a space, used to identify read and unread messages.

SpaceUpdatedEventData

Event payload for an updated space.

Thread

A thread in a Google Chat space. For example usage, see Start or reply to a message thread.

ThreadReadState

A user’s read state within a thread, used to identify read and unread messages.

UpdateMembershipRequest

Request message for updating a membership.

UpdateMessageRequest

Request to update a message.

UpdateSpaceReadStateRequest

Request message for UpdateSpaceReadState API.

UpdateSpaceRequest

A request to update a single space.

UploadAttachmentRequest

Request to upload an attachment.

UploadAttachmentResponse

Response of uploading an attachment.

User

A user in Google Chat. When returned as an output from a request, if your Chat app authenticates as a user, the output for a User resource only populates the user’s name and type.

UserMentionMetadata

Annotation metadata for user mentions (@).

WidgetMarkup

A widget is a UI element that presents text and images.

Enums

AnnotationType

Type of the annotation.

HistoryState

The history state for messages and spaces. Specifies how long messages and conversation threads are kept after creation.