Module google_api_proto::google::chat::v1::widget_markup

Nested message and enum types in WidgetMarkup.

Modules

button

Nested message and enum types in Button.

form_action

Nested message and enum types in FormAction.

image_button

Nested message and enum types in ImageButton.

key_value

Nested message and enum types in KeyValue.

on_click

Nested message and enum types in OnClick.

Structs

Button

A button. Can be a text button or an image button.

FormAction

A form action describes the behavior when the form is submitted. For example, you can invoke Apps Script to handle the form.

Image

An image that’s specified by a URL and can have an onclick action.

ImageButton

An image button with an onclick action.

KeyValue

A UI element contains a key (label) and a value (content). This element can also contain some actions such as onclick button.

OnClick

An onclick action (for example, open a link).

OpenLink

A link that opens a new window.

TextButton

A button with text and onclick action.

TextParagraph

A paragraph of text. Formatted text supported. For more information about formatting text, see Formatting text in Google Chat apps and Formatting text in Google Workspace Add-ons.

Enums

Data

A WidgetMarkup can only have one of the following items. You can use multiple WidgetMarkup fields to display more items.

Icon

The set of supported icons.