Module google_api_proto::google::cloud

Modules

abuseevent

accessapproval

advisorynotifications

aiplatform

alloydb

apigateway

apigeeconnect

apigeeregistry

apihub

apphub

asset

assuredworkloads

audit

automl

backupdr

baremetalsolution

batch

beyondcorp

bigquery

billing

binaryauthorization

blockchainnodeengine

certificatemanager

channel

cloudcontrolspartner

clouddms

cloudsetup

commerce

common

compute

confidentialcomputing

config

connectors

contactcenterinsights

contentwarehouse

datacatalog

dataform

datafusion

datalabeling

datapipelines

dataplex

dataproc

dataqna

datastream

deploy

developerconnect

dialogflow

discoveryengine

documentai

domains

edgecontainer

edgenetwork

enterpriseknowledgegraph

essentialcontacts

eventarc

filestore

functions

gdchardwaremanagement

gkebackup

gkeconnect

gkehub

gkemulticloud

gsuiteaddons

healthcare

iap

identitytoolkit

ids

integrations

iot

kms

language

lifesciences

location

managedidentities

managedkafka

mediatranslation

memcache

metastore

migrationcenter

netapp

networkanalyzer

networkconnectivity

networkmanagement

networksecurity

networkservices

notebooks

optimization

orchestration

orgpolicy

osconfig

oslogin

parallelstore

paymentgateway

phishingprotection

policysimulator

policytroubleshooter

privatecatalog

privilegedaccessmanager

pubsublite

rapidmigrationassessment

recaptchaenterprise

recommendationengine

recommender

redis

resourcemanager

resourcesettings

retail

run

runtimeconfig

saasaccelerator

scheduler

secretmanager

secrets

securesourcemanager

security

securitycenter

securitycentermanagement

securityposture

sensitiveaction

servicedirectory

servicehealth

shell

speech

sql

storageinsights

stream

support

talent

tasks

telcoautomation

texttospeech

timeseriesinsights

tpu

translation

video

videointelligence

vision

visionai

vmmigration

vmwareengine

vpcaccess

webrisk

websecurityscanner

workflows

workstations

Enums

OperationResponseMapping

An enum to be used to mark the essential (for polling) fields in an API-specific Operation object. A custom Operation object may contain many different fields, but only few of them are essential to conduct a successful polling process.