Module google_api_proto::google::cloud::abuseevent::logging::v1

Modules

abuse_event

Nested message and enum types in AbuseEvent.

leaked_credential_event

Nested message and enum types in LeakedCredentialEvent.

Structs

AbuseEvent

Logged event discussing an abuse finding on the monitored resource.

ApiKeyCredential

Information about leaked API keys.

CryptoMiningEvent

Information about a cryptocurrency mining event observed on the monitored resource.

DecisionEscalationEvent

Information about the escalation of enforcement action on the monitored resource. Eg: Used for repeat offenders.

HarmfulContentEvent

Information about the phishing/malware URIs associated with the monitored resource.

LeakedCredentialEvent

Information about the leaked client credentials observed on the monitored resource.

ReinstatementEvent

Information about the reinstatement issued on the monitored resource.

ServiceAccountCredential

Information about leaked service accounts.