Struct google_api_proto::google::cloud::accessapproval::v1::access_approval_client::AccessApprovalClient

pub struct AccessApprovalClient<T> { /* private fields */ }

This API allows a customer to manage accesses to cloud resources by Google personnel. It defines the following resource model:

The API has a collection of [ApprovalRequest][google.cloud.accessapproval.v1.ApprovalRequest] resources, named approvalRequests/{approval_request}

The API has top-level settings per Project/Folder/Organization, named accessApprovalSettings

The service also periodically emails a list of recipients, defined at the Project/Folder/Organization level in the accessApprovalSettings, when there is a pending ApprovalRequest for them to act on. The ApprovalRequests can also optionally be published to a Pub/Sub topic owned by the customer (contact support if you would like to enable Pub/Sub notifications).

ApprovalRequests can be approved or dismissed. Google personnel can only access the indicated resource or resources if the request is approved (subject to some exclusions: https://cloud.google.com/access-approval/docs/overview#exclusions).

Note: Using Access Approval functionality will mean that Google may not be able to meet the SLAs for your chosen products, as any support response times may be dramatically increased. As such the SLAs do not apply to any service disruption to the extent impacted by Customer’s use of Access Approval. Do not enable Access Approval for projects where you may require high service availability and rapid response by Google Cloud Support.

After a request is approved or dismissed, no further action may be taken on it. Requests with the requested_expiration in the past or with no activity for 14 days are considered dismissed. When an approval expires, the request is considered dismissed.

If a request is not approved or dismissed, we call it pending.

Implementations

impl<T> AccessApprovalClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AccessApprovalClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_approval_requests( &mut self, request: impl IntoRequest<ListApprovalRequestsMessage> ) -> Result<Response<ListApprovalRequestsResponse>, Status>

Lists approval requests associated with a project, folder, or organization. Approval requests can be filtered by state (pending, active, dismissed). The order is reverse chronological.

pub async fn get_approval_request( &mut self, request: impl IntoRequest<GetApprovalRequestMessage> ) -> Result<Response<ApprovalRequest>, Status>

Gets an approval request. Returns NOT_FOUND if the request does not exist.

pub async fn approve_approval_request( &mut self, request: impl IntoRequest<ApproveApprovalRequestMessage> ) -> Result<Response<ApprovalRequest>, Status>

Approves a request and returns the updated ApprovalRequest.

Returns NOT_FOUND if the request does not exist. Returns FAILED_PRECONDITION if the request exists but is not in a pending state.

pub async fn dismiss_approval_request( &mut self, request: impl IntoRequest<DismissApprovalRequestMessage> ) -> Result<Response<ApprovalRequest>, Status>

Dismisses a request. Returns the updated ApprovalRequest.

NOTE: This does not deny access to the resource if another request has been made and approved. It is equivalent in effect to ignoring the request altogether.

Returns NOT_FOUND if the request does not exist.

Returns FAILED_PRECONDITION if the request exists but is not in a pending state.

pub async fn invalidate_approval_request( &mut self, request: impl IntoRequest<InvalidateApprovalRequestMessage> ) -> Result<Response<ApprovalRequest>, Status>

Invalidates an existing ApprovalRequest. Returns the updated ApprovalRequest.

NOTE: This does not deny access to the resource if another request has been made and approved. It only invalidates a single approval.

Returns FAILED_PRECONDITION if the request exists but is not in an approved state.

pub async fn get_access_approval_settings( &mut self, request: impl IntoRequest<GetAccessApprovalSettingsMessage> ) -> Result<Response<AccessApprovalSettings>, Status>

Gets the settings associated with a project, folder, or organization.

pub async fn update_access_approval_settings( &mut self, request: impl IntoRequest<UpdateAccessApprovalSettingsMessage> ) -> Result<Response<AccessApprovalSettings>, Status>

Updates the settings associated with a project, folder, or organization. Settings to update are determined by the value of field_mask.

pub async fn delete_access_approval_settings( &mut self, request: impl IntoRequest<DeleteAccessApprovalSettingsMessage> ) -> Result<Response<()>, Status>

Deletes the settings associated with a project, folder, or organization. This will have the effect of disabling Access Approval for the project, folder, or organization, but only if all ancestors also have Access Approval disabled. If Access Approval is enabled at a higher level of the hierarchy, then Access Approval will still be enabled at this level as the settings are inherited.

pub async fn get_access_approval_service_account( &mut self, request: impl IntoRequest<GetAccessApprovalServiceAccountMessage> ) -> Result<Response<AccessApprovalServiceAccount>, Status>

Retrieves the service account that is used by Access Approval to access KMS keys for signing approved approval requests.