Module google_api_proto::google::cloud::advisorynotifications::v1

Modules

advisory_notifications_service_client

Generated client implementations.

attachment

Nested message and enum types in Attachment.

csv

Nested message and enum types in Csv.

message

Nested message and enum types in Message.

Structs

Attachment

Attachment with specific information about the issue.

Csv

A representation of a CSV file attachment, as a list of column headers and a list of data rows.

GetNotificationRequest

Request for fetching a notification.

GetSettingsRequest

Request of GetSettings endpoint.

ListNotificationsRequest

Request for fetching all notifications for a given parent.

ListNotificationsResponse

Response of ListNotifications endpoint.

Message

A message which contains notification details.

Notification

A notification object for notifying customers about security and privacy issues.

NotificationSettings

Settings for each NotificationType.

Settings

Settings for Advisory Notifications.

Subject

A subject line of a notification.

Text

A text object containing the English text and its localized copies.

UpdateSettingsRequest

Request of UpdateSettings endpoint.

Enums

LocalizationState

Status of localized text.

NotificationType

Type of notification

NotificationView

Notification view.