Module google_api_proto::google::cloud::aiplatform::v1::batch_prediction_job

Nested message and enum types in BatchPredictionJob.

Modules

input_config

Nested message and enum types in InputConfig.

output_config

Nested message and enum types in OutputConfig.

output_info

Nested message and enum types in OutputInfo.

Structs

InputConfig

Configures the input to [BatchPredictionJob][google.cloud.aiplatform.v1.BatchPredictionJob]. See [Model.supported_input_storage_formats][google.cloud.aiplatform.v1.Model.supported_input_storage_formats] for Model’s supported input formats, and how instances should be expressed via any of them.

OutputConfig

Configures the output of [BatchPredictionJob][google.cloud.aiplatform.v1.BatchPredictionJob]. See [Model.supported_output_storage_formats][google.cloud.aiplatform.v1.Model.supported_output_storage_formats] for supported output formats, and how predictions are expressed via any of them.

OutputInfo

Further describes this job’s output. Supplements [output_config][google.cloud.aiplatform.v1.BatchPredictionJob.output_config].