Struct google_api_proto::google::cloud::aiplatform::v1::featurestore_service_client::FeaturestoreServiceClient

pub struct FeaturestoreServiceClient<T> { /* private fields */ }

The service that handles CRUD and List for resources for Featurestore.

Implementations

impl<T> FeaturestoreServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> FeaturestoreServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_featurestore( &mut self, request: impl IntoRequest<CreateFeaturestoreRequest> ) -> Result<Response<Operation>, Status>

Creates a new Featurestore in a given project and location.

pub async fn get_featurestore( &mut self, request: impl IntoRequest<GetFeaturestoreRequest> ) -> Result<Response<Featurestore>, Status>

Gets details of a single Featurestore.

pub async fn list_featurestores( &mut self, request: impl IntoRequest<ListFeaturestoresRequest> ) -> Result<Response<ListFeaturestoresResponse>, Status>

Lists Featurestores in a given project and location.

pub async fn update_featurestore( &mut self, request: impl IntoRequest<UpdateFeaturestoreRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Featurestore.

pub async fn delete_featurestore( &mut self, request: impl IntoRequest<DeleteFeaturestoreRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Featurestore. The Featurestore must not contain any EntityTypes or force must be set to true for the request to succeed.

pub async fn create_entity_type( &mut self, request: impl IntoRequest<CreateEntityTypeRequest> ) -> Result<Response<Operation>, Status>

Creates a new EntityType in a given Featurestore.

pub async fn get_entity_type( &mut self, request: impl IntoRequest<GetEntityTypeRequest> ) -> Result<Response<EntityType>, Status>

Gets details of a single EntityType.

pub async fn list_entity_types( &mut self, request: impl IntoRequest<ListEntityTypesRequest> ) -> Result<Response<ListEntityTypesResponse>, Status>

Lists EntityTypes in a given Featurestore.

pub async fn update_entity_type( &mut self, request: impl IntoRequest<UpdateEntityTypeRequest> ) -> Result<Response<EntityType>, Status>

Updates the parameters of a single EntityType.

pub async fn delete_entity_type( &mut self, request: impl IntoRequest<DeleteEntityTypeRequest> ) -> Result<Response<Operation>, Status>

Deletes a single EntityType. The EntityType must not have any Features or force must be set to true for the request to succeed.

pub async fn create_feature( &mut self, request: impl IntoRequest<CreateFeatureRequest> ) -> Result<Response<Operation>, Status>

Creates a new Feature in a given EntityType.

pub async fn batch_create_features( &mut self, request: impl IntoRequest<BatchCreateFeaturesRequest> ) -> Result<Response<Operation>, Status>

Creates a batch of Features in a given EntityType.

pub async fn get_feature( &mut self, request: impl IntoRequest<GetFeatureRequest> ) -> Result<Response<Feature>, Status>

Gets details of a single Feature.

pub async fn list_features( &mut self, request: impl IntoRequest<ListFeaturesRequest> ) -> Result<Response<ListFeaturesResponse>, Status>

Lists Features in a given EntityType.

pub async fn update_feature( &mut self, request: impl IntoRequest<UpdateFeatureRequest> ) -> Result<Response<Feature>, Status>

Updates the parameters of a single Feature.

pub async fn delete_feature( &mut self, request: impl IntoRequest<DeleteFeatureRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Feature.

pub async fn import_feature_values( &mut self, request: impl IntoRequest<ImportFeatureValuesRequest> ) -> Result<Response<Operation>, Status>

Imports Feature values into the Featurestore from a source storage.

The progress of the import is tracked by the returned operation. The imported features are guaranteed to be visible to subsequent read operations after the operation is marked as successfully done.

If an import operation fails, the Feature values returned from reads and exports may be inconsistent. If consistency is required, the caller must retry the same import request again and wait till the new operation returned is marked as successfully done.

There are also scenarios where the caller can cause inconsistency.

Source data for import contains multiple distinct Feature values for the same entity ID and timestamp.

Source is modified during an import. This includes adding, updating, or removing source data and/or metadata. Examples of updating metadata include but are not limited to changing storage location, storage class, or retention policy.

Online serving cluster is under-provisioned.

pub async fn batch_read_feature_values( &mut self, request: impl IntoRequest<BatchReadFeatureValuesRequest> ) -> Result<Response<Operation>, Status>

Batch reads Feature values from a Featurestore.

This API enables batch reading Feature values, where each read instance in the batch may read Feature values of entities from one or more EntityTypes. Point-in-time correctness is guaranteed for Feature values of each read instance as of each instance’s read timestamp.

pub async fn export_feature_values( &mut self, request: impl IntoRequest<ExportFeatureValuesRequest> ) -> Result<Response<Operation>, Status>

Exports Feature values from all the entities of a target EntityType.

pub async fn delete_feature_values( &mut self, request: impl IntoRequest<DeleteFeatureValuesRequest> ) -> Result<Response<Operation>, Status>

Delete Feature values from Featurestore.

The progress of the deletion is tracked by the returned operation. The deleted feature values are guaranteed to be invisible to subsequent read operations after the operation is marked as successfully done.

If a delete feature values operation fails, the feature values returned from reads and exports may be inconsistent. If consistency is required, the caller must retry the same delete request again and wait till the new operation returned is marked as successfully done.

pub async fn search_features( &mut self, request: impl IntoRequest<SearchFeaturesRequest> ) -> Result<Response<SearchFeaturesResponse>, Status>

Searches Features matching a query in a given project.