Module google_api_proto::google::cloud::aiplatform::v1beta1::featurestore_monitoring_config

Nested message and enum types in FeaturestoreMonitoringConfig.

Modules

import_features_analysis

Nested message and enum types in ImportFeaturesAnalysis.

threshold_config

Nested message and enum types in ThresholdConfig.

Structs

ImportFeaturesAnalysis

Configuration of the Featurestore’s ImportFeature Analysis Based Monitoring. This type of analysis generates statistics for values of each Feature imported by every [ImportFeatureValues][google.cloud.aiplatform.v1beta1.FeaturestoreService.ImportFeatureValues] operation.

SnapshotAnalysis

Configuration of the Featurestore’s Snapshot Analysis Based Monitoring. This type of analysis generates statistics for each Feature based on a snapshot of the latest feature value of each entities every monitoring_interval.

ThresholdConfig

The config for Featurestore Monitoring threshold.