Struct google_api_proto::google::cloud::aiplatform::v1beta1::DeployedIndex

pub struct DeployedIndex {
    pub id: String,
    pub index: String,
    pub display_name: String,
    pub create_time: Option<Timestamp>,
    pub private_endpoints: Option<IndexPrivateEndpoints>,
    pub index_sync_time: Option<Timestamp>,
    pub automatic_resources: Option<AutomaticResources>,
    pub dedicated_resources: Option<DedicatedResources>,
    pub enable_access_logging: bool,
    pub deployed_index_auth_config: Option<DeployedIndexAuthConfig>,
    pub reserved_ip_ranges: Vec<String>,
    pub deployment_group: String,
}

A deployment of an Index. IndexEndpoints contain one or more DeployedIndexes.

Fields
id: String

Required. The user specified ID of the DeployedIndex. The ID can be up to 128 characters long and must start with a letter and only contain letters, numbers, and underscores. The ID must be unique within the project it is created in.

index: String

Required. The name of the Index this is the deployment of. We may refer to this Index as the DeployedIndex’s “original” Index.

display_name: String

The display name of the DeployedIndex. If not provided upon creation, the Index’s display_name is used.

create_time: Option<Timestamp>

Output only. Timestamp when the DeployedIndex was created.

private_endpoints: Option<IndexPrivateEndpoints>

Output only. Provides paths for users to send requests directly to the deployed index services running on Cloud via private services access. This field is populated if [network][google.cloud.aiplatform.v1beta1.IndexEndpoint.network] is configured.

index_sync_time: Option<Timestamp>

Output only. The DeployedIndex may depend on various data on its original Index. Additionally when certain changes to the original Index are being done (e.g. when what the Index contains is being changed) the DeployedIndex may be asynchronously updated in the background to reflect these changes. If this timestamp’s value is at least the [Index.update_time][google.cloud.aiplatform.v1beta1.Index.update_time] of the original Index, it means that this DeployedIndex and the original Index are in sync. If this timestamp is older, then to see which updates this DeployedIndex already contains (and which it does not), one must [list][google.longrunning.Operations.ListOperations] the operations that are running on the original Index. Only the successfully completed Operations with [update_time][google.cloud.aiplatform.v1beta1.GenericOperationMetadata.update_time] equal or before this sync time are contained in this DeployedIndex.

automatic_resources: Option<AutomaticResources>

Optional. A description of resources that the DeployedIndex uses, which to large degree are decided by Vertex AI, and optionally allows only a modest additional configuration. If min_replica_count is not set, the default value is 2 (we don’t provide SLA when min_replica_count=1). If max_replica_count is not set, the default value is min_replica_count. The max allowed replica count is 1000.

dedicated_resources: Option<DedicatedResources>

Optional. A description of resources that are dedicated to the DeployedIndex, and that need a higher degree of manual configuration. The field min_replica_count must be set to a value strictly greater than 0, or else validation will fail. We don’t provide SLA when min_replica_count=1. If max_replica_count is not set, the default value is min_replica_count. The max allowed replica count is 1000.

Available machine types for SMALL shard: e2-standard-2 and all machine types available for MEDIUM and LARGE shard.

Available machine types for MEDIUM shard: e2-standard-16 and all machine types available for LARGE shard.

Available machine types for LARGE shard: e2-highmem-16, n2d-standard-32.

n1-standard-16 and n1-standard-32 are still available, but we recommend e2-standard-16 and e2-highmem-16 for cost efficiency.

enable_access_logging: bool

Optional. If true, private endpoint’s access logs are sent to Cloud Logging.

These logs are like standard server access logs, containing information like timestamp and latency for each MatchRequest.

Note that logs may incur a cost, especially if the deployed index receives a high queries per second rate (QPS). Estimate your costs before enabling this option.

deployed_index_auth_config: Option<DeployedIndexAuthConfig>

Optional. If set, the authentication is enabled for the private endpoint.

reserved_ip_ranges: Vec<String>

Optional. A list of reserved ip ranges under the VPC network that can be used for this DeployedIndex.

If set, we will deploy the index within the provided ip ranges. Otherwise, the index might be deployed to any ip ranges under the provided VPC network.

The value should be the name of the address (https://cloud.google.com/compute/docs/reference/rest/v1/addresses) Example: [‘vertex-ai-ip-range’].

For more information about subnets and network IP ranges, please see https://cloud.google.com/vpc/docs/subnets#manually_created_subnet_ip_ranges.

deployment_group: String

Optional. The deployment group can be no longer than 64 characters (eg: ‘test’, ‘prod’). If not set, we will use the ‘default’ deployment group.

Creating deployment_groups with reserved_ip_ranges is a recommended practice when the peered network has multiple peering ranges. This creates your deployments from predictable IP spaces for easier traffic administration. Also, one deployment_group (except ‘default’) can only be used with the same reserved_ip_ranges which means if the deployment_group has been used with reserved_ip_ranges: [a, b, c], using it with [a, b] or [d, e] is disallowed.

Note: we only support up to 5 deployment groups(not including ‘default’).