Struct google_api_proto::google::cloud::aiplatform::v1beta1::PipelineJob

pub struct PipelineJob {

    pub name: String,
    pub display_name: String,
    pub create_time: Option<Timestamp>,
    pub start_time: Option<Timestamp>,
    pub end_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub pipeline_spec: Option<Struct>,
    pub state: i32,
    pub job_detail: Option<PipelineJobDetail>,
    pub error: Option<Status>,
    pub labels: BTreeMap<String, String>,
    pub runtime_config: Option<RuntimeConfig>,
    pub encryption_spec: Option<EncryptionSpec>,
    pub service_account: String,
    pub network: String,
    pub reserved_ip_ranges: Vec<String>,
    pub template_uri: String,
    pub template_metadata: Option<PipelineTemplateMetadata>,
    pub schedule_name: String,
    pub preflight_validations: bool,
    pub satisfies_pzs: bool,
    pub satisfies_pzi: bool,

}

An instance of a machine learning PipelineJob.

Fields
name: String

Output only. The resource name of the PipelineJob.

display_name: String

The display name of the Pipeline. The name can be up to 128 characters long and can consist of any UTF-8 characters.

create_time: Option<Timestamp>

Output only. Pipeline creation time.

start_time: Option<Timestamp>

Output only. Pipeline start time.

end_time: Option<Timestamp>

Output only. Pipeline end time.

update_time: Option<Timestamp>

Output only. Timestamp when this PipelineJob was most recently updated.

pipeline_spec: Option<Struct>

The spec of the pipeline.

state: i32

Output only. The detailed state of the job.

job_detail: Option<PipelineJobDetail>

Output only. The details of pipeline run. Not available in the list view.

error: Option<Status>

Output only. The error that occurred during pipeline execution. Only populated when the pipeline’s state is FAILED or CANCELLED.

labels: BTreeMap<String, String>

The labels with user-defined metadata to organize PipelineJob.

Label keys and values can be no longer than 64 characters (Unicode codepoints), can only contain lowercase letters, numeric characters, underscores and dashes. International characters are allowed.

See https://goo.gl/xmQnxf for more information and examples of labels.

Note there is some reserved label key for Vertex AI Pipelines.

vertex-ai-pipelines-run-billing-id, user set value will get overrided.

runtime_config: Option<RuntimeConfig>

Runtime config of the pipeline.

encryption_spec: Option<EncryptionSpec>

Customer-managed encryption key spec for a pipelineJob. If set, this PipelineJob and all of its sub-resources will be secured by this key.

service_account: String

The service account that the pipeline workload runs as. If not specified, the Compute Engine default service account in the project will be used. See https://cloud.google.com/compute/docs/access/service-accounts#default_service_account

Users starting the pipeline must have the iam.serviceAccounts.actAs permission on this service account.

network: String

The full name of the Compute Engine network to which the Pipeline Job’s workload should be peered. For example, projects/12345/global/networks/myVPC. Format is of the form projects/{project}/global/networks/{network}. Where {project} is a project number, as in 12345, and {network} is a network name.

Private services access must already be configured for the network. Pipeline job will apply the network configuration to the Google Cloud resources being launched, if applied, such as Vertex AI Training or Dataflow job. If left unspecified, the workload is not peered with any network.

reserved_ip_ranges: Vec<String>

A list of names for the reserved ip ranges under the VPC network that can be used for this Pipeline Job’s workload.

If set, we will deploy the Pipeline Job’s workload within the provided ip ranges. Otherwise, the job will be deployed to any ip ranges under the provided VPC network.

Example: [‘vertex-ai-ip-range’].

template_uri: String

A template uri from where the [PipelineJob.pipeline_spec][google.cloud.aiplatform.v1beta1.PipelineJob.pipeline_spec], if empty, will be downloaded. Currently, only uri from Vertex Template Registry & Gallery is supported. Reference to https://cloud.google.com/vertex-ai/docs/pipelines/create-pipeline-template.

template_metadata: Option<PipelineTemplateMetadata>

Output only. Pipeline template metadata. Will fill up fields if [PipelineJob.template_uri][google.cloud.aiplatform.v1beta1.PipelineJob.template_uri] is from supported template registry.

schedule_name: String

Output only. The schedule resource name. Only returned if the Pipeline is created by Schedule API.

preflight_validations: bool

Optional. Whether to do component level validations before job creation.

satisfies_pzs: bool

Output only. Reserved for future use.

satisfies_pzi: bool

Output only. Reserved for future use.

Implementations

impl PipelineJob

pub fn state(&self) -> PipelineState

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: PipelineState)

Sets state to the provided enum value.