Struct google_api_proto::google::cloud::aiplatform::v1beta1::SearchDataItemsRequest

pub struct SearchDataItemsRequest {
    pub dataset: String,
    pub saved_query: String,
    pub data_labeling_job: String,
    pub data_item_filter: String,
    pub annotations_filter: String,
    pub annotation_filters: Vec<String>,
    pub field_mask: Option<FieldMask>,
    pub annotations_limit: i32,
    pub page_size: i32,
    pub order_by: String,
    pub page_token: String,
    pub order: Option<Order>,
}

Request message for [DatasetService.SearchDataItems][google.cloud.aiplatform.v1beta1.DatasetService.SearchDataItems].

Fields
dataset: String

Required. The resource name of the Dataset from which to search DataItems. Format: projects/{project}/locations/{location}/datasets/{dataset}

saved_query: String
๐Deprecated

The resource name of a SavedQuery(annotation set in UI). Format: projects/{project}/locations/{location}/datasets/{dataset}/savedQueries/{saved_query} All of the search will be done in the context of this SavedQuery.

data_labeling_job: String

The resource name of a DataLabelingJob. Format: projects/{project}/locations/{location}/dataLabelingJobs/{data_labeling_job} If this field is set, all of the search will be done in the context of this DataLabelingJob.

data_item_filter: String

An expression for filtering the DataItem that will be returned.

data_item_id - for = or !=.

labeled - for = or !=.

has_annotation(ANNOTATION_SPEC_ID) - true only for DataItem that have at least one annotation with annotation_spec_id = ANNOTATION_SPEC_ID in the context of SavedQuery or DataLabelingJob.

For example:

data_item=1

has_annotation(5)

annotations_filter: String
๐Deprecated

An expression for filtering the Annotations that will be returned per DataItem.

annotation_spec_id - for = or !=.

annotation_filters: Vec<String>

An expression that specifies what Annotations will be returned per DataItem. Annotations satisfied either of the conditions will be returned.

annotation_spec_id - for = or !=. Must specify saved_query_id= - saved query id that annotations should belong to.

field_mask: Option<FieldMask>

Mask specifying which fields of [DataItemView][google.cloud.aiplatform.v1beta1.DataItemView] to read.

annotations_limit: i32

If set, only up to this many of Annotations will be returned per DataItemView. The maximum value is 1000. If not set, the maximum value will be used.

page_size: i32

Requested page size. Server may return fewer results than requested. Default and maximum page size is 100.

order_by: String
๐Deprecated

A comma-separated list of fields to order by, sorted in ascending order. Use โdescโ after a field name for descending.

page_token: String

A token identifying a page of results for the server to return Typically obtained via [SearchDataItemsResponse.next_page_token][google.cloud.aiplatform.v1beta1.SearchDataItemsResponse.next_page_token] of the previous [DatasetService.SearchDataItems][google.cloud.aiplatform.v1beta1.DatasetService.SearchDataItems] call.

order: Option<Order>