Struct google_api_proto::google::cloud::alloydb::v1::alloy_db_admin_client::AlloyDbAdminClient

pub struct AlloyDbAdminClient<T> { /* private fields */ }

Service describing handlers for resources

Implementations

impl<T> AlloyDbAdminClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AlloyDbAdminClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_clusters( &mut self, request: impl IntoRequest<ListClustersRequest> ) -> Result<Response<ListClustersResponse>, Status>

Lists Clusters in a given project and location.

pub async fn get_cluster( &mut self, request: impl IntoRequest<GetClusterRequest> ) -> Result<Response<Cluster>, Status>

Gets details of a single Cluster.

pub async fn create_cluster( &mut self, request: impl IntoRequest<CreateClusterRequest> ) -> Result<Response<Operation>, Status>

Creates a new Cluster in a given project and location.

pub async fn update_cluster( &mut self, request: impl IntoRequest<UpdateClusterRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Cluster.

pub async fn delete_cluster( &mut self, request: impl IntoRequest<DeleteClusterRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Cluster.

pub async fn promote_cluster( &mut self, request: impl IntoRequest<PromoteClusterRequest> ) -> Result<Response<Operation>, Status>

Promotes a SECONDARY cluster. This turns down replication from the PRIMARY cluster and promotes a secondary cluster into its own standalone cluster. Imperative only.

pub async fn restore_cluster( &mut self, request: impl IntoRequest<RestoreClusterRequest> ) -> Result<Response<Operation>, Status>

Creates a new Cluster in a given project and location, with a volume restored from the provided source, either a backup ID or a point-in-time and a source cluster.

pub async fn create_secondary_cluster( &mut self, request: impl IntoRequest<CreateSecondaryClusterRequest> ) -> Result<Response<Operation>, Status>

Creates a cluster of type SECONDARY in the given location using the primary cluster as the source.

pub async fn list_instances( &mut self, request: impl IntoRequest<ListInstancesRequest> ) -> Result<Response<ListInstancesResponse>, Status>

Lists Instances in a given project and location.

pub async fn get_instance( &mut self, request: impl IntoRequest<GetInstanceRequest> ) -> Result<Response<Instance>, Status>

Gets details of a single Instance.

pub async fn create_instance( &mut self, request: impl IntoRequest<CreateInstanceRequest> ) -> Result<Response<Operation>, Status>

Creates a new Instance in a given project and location.

pub async fn create_secondary_instance( &mut self, request: impl IntoRequest<CreateSecondaryInstanceRequest> ) -> Result<Response<Operation>, Status>

Creates a new SECONDARY Instance in a given project and location.

pub async fn batch_create_instances( &mut self, request: impl IntoRequest<BatchCreateInstancesRequest> ) -> Result<Response<Operation>, Status>

Creates new instances under the given project, location and cluster. There can be only one primary instance in a cluster. If the primary instance exists in the cluster as well as this request, then API will throw an error. The primary instance should exist before any read pool instance is created. If the primary instance is a part of the request payload, then the API will take care of creating instances in the correct order. This method is here to support Google-internal use cases, and is not meant for external customers to consume. Please do not start relying on it; its behavior is subject to change without notice.

pub async fn update_instance( &mut self, request: impl IntoRequest<UpdateInstanceRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Instance.

pub async fn delete_instance( &mut self, request: impl IntoRequest<DeleteInstanceRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Instance.

pub async fn failover_instance( &mut self, request: impl IntoRequest<FailoverInstanceRequest> ) -> Result<Response<Operation>, Status>

Forces a Failover for a highly available instance. Failover promotes the HA standby instance as the new primary. Imperative only.

pub async fn inject_fault( &mut self, request: impl IntoRequest<InjectFaultRequest> ) -> Result<Response<Operation>, Status>

Injects fault in an instance. Imperative only.

pub async fn restart_instance( &mut self, request: impl IntoRequest<RestartInstanceRequest> ) -> Result<Response<Operation>, Status>

Restart an Instance in a cluster. Imperative only.

pub async fn list_backups( &mut self, request: impl IntoRequest<ListBackupsRequest> ) -> Result<Response<ListBackupsResponse>, Status>

Lists Backups in a given project and location.

pub async fn get_backup( &mut self, request: impl IntoRequest<GetBackupRequest> ) -> Result<Response<Backup>, Status>

Gets details of a single Backup.

pub async fn create_backup( &mut self, request: impl IntoRequest<CreateBackupRequest> ) -> Result<Response<Operation>, Status>

Creates a new Backup in a given project and location.

pub async fn update_backup( &mut self, request: impl IntoRequest<UpdateBackupRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Backup.

pub async fn delete_backup( &mut self, request: impl IntoRequest<DeleteBackupRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Backup.

pub async fn list_supported_database_flags( &mut self, request: impl IntoRequest<ListSupportedDatabaseFlagsRequest> ) -> Result<Response<ListSupportedDatabaseFlagsResponse>, Status>

Lists SupportedDatabaseFlags for a given project and location.

pub async fn list_users( &mut self, request: impl IntoRequest<ListUsersRequest> ) -> Result<Response<ListUsersResponse>, Status>

Lists Users in a given project and location.

pub async fn get_user( &mut self, request: impl IntoRequest<GetUserRequest> ) -> Result<Response<User>, Status>

Gets details of a single User.

pub async fn create_user( &mut self, request: impl IntoRequest<CreateUserRequest> ) -> Result<Response<User>, Status>

Creates a new User in a given project, location, and cluster.

pub async fn update_user( &mut self, request: impl IntoRequest<UpdateUserRequest> ) -> Result<Response<User>, Status>

Updates the parameters of a single User.

pub async fn delete_user( &mut self, request: impl IntoRequest<DeleteUserRequest> ) -> Result<Response<()>, Status>

Deletes a single User.