Module google_api_proto::google::cloud::alloydb::v1alpha

Modules

alloy_db_admin_client

Generated client implementations.

automated_backup_policy

Nested message and enum types in AutomatedBackupPolicy.

backup

Nested message and enum types in Backup.

batch_create_instance_status

Nested message and enum types in BatchCreateInstanceStatus.

cluster

Nested message and enum types in Cluster.

encryption_info

Nested message and enum types in EncryptionInfo.

inject_fault_request

Nested message and enum types in InjectFaultRequest.

instance

Nested message and enum types in Instance.

migration_source

Nested message and enum types in MigrationSource.

operation_metadata

Nested message and enum types in OperationMetadata.

restore_cluster_request

Nested message and enum types in RestoreClusterRequest.

ssl_config

Nested message and enum types in SslConfig.

supported_database_flag

Nested message and enum types in SupportedDatabaseFlag.

user

Nested message and enum types in User.

Structs

AutomatedBackupPolicy

Message describing the user-specified automated backup policy.

Backup

Message describing Backup object

BackupSource

Message describing a BackupSource.

BatchCreateInstanceStatus

Message for current status of an instance in the BatchCreateInstances operation. For example, lets say a BatchCreateInstances workflow has 4 instances, Instance1 through Instance4. Lets also assume that 2 instances succeeded but the third failed to create and the 4th was never picked up for creation because of failure of the previous one. Then, resulting states would look something like:

BatchCreateInstancesMetadata

Message for metadata that is specific to BatchCreateInstances API. NEXT_ID: 3

BatchCreateInstancesRequest

Message for creating a batch of instances under the specified cluster.

BatchCreateInstancesResponse

Message for creating batches of instances in a cluster.

Cluster

A cluster is a collection of regional AlloyDB resources. It can include a primary instance and one or more read pool instances. All cluster resources share a storage layer, which scales as needed.

ConnectionInfo

ConnectionInfo singleton resource. https://google.aip.dev/156

ContinuousBackupConfig

ContinuousBackupConfig describes the continuous backups recovery configurations of a cluster.

ContinuousBackupInfo

ContinuousBackupInfo describes the continuous backup properties of a cluster.

ContinuousBackupSource

Message describing a ContinuousBackupSource.

CreateBackupRequest

Message for creating a Backup

CreateClusterRequest

Message for creating a Cluster

CreateInstanceRequest

Message for creating a Instance

CreateInstanceRequests

See usage below for notes.

CreateSecondaryClusterRequest

CreateSecondaryInstanceRequest

Message for creating a Secondary Instance

CreateUserRequest

Message for creating a User

Database

Message describing Database object.

DeleteBackupRequest

Message for deleting a Backup

DeleteClusterRequest

Message for deleting a Cluster

DeleteInstanceRequest

Message for deleting a Instance

DeleteUserRequest

Message for deleting a User

EncryptionConfig

EncryptionConfig describes the encryption config of a cluster or a backup that is encrypted with a CMEK (customer-managed encryption key).

EncryptionInfo

EncryptionInfo describes the encryption information of a cluster or a backup.

FailoverInstanceRequest

Message for triggering failover on an Instance

GenerateClientCertificateRequest

Message for requests to generate a client certificate signed by the Cluster CA.

GenerateClientCertificateResponse

Message returned by a GenerateClientCertificate operation.

GetBackupRequest

Message for getting a Backup

GetClusterRequest

Message for getting a Cluster

GetConnectionInfoRequest

Request message for GetConnectionInfo.

GetInstanceRequest

Message for getting a Instance

GetUserRequest

Message for getting a User

InjectFaultRequest

Message for triggering fault injection on an instance

Instance

An Instance is a computing unit that an end customer can connect to. It’s the main unit of computing resources in AlloyDB.

ListBackupsRequest

Message for requesting list of Backups

ListBackupsResponse

Message for response to listing Backups

ListClustersRequest

Message for requesting list of Clusters

ListClustersResponse

Message for response to listing Clusters

ListDatabasesRequest

Message for requesting list of Databases.

ListDatabasesResponse

Message for response to listing Databases.

ListInstancesRequest

Message for requesting list of Instances

ListInstancesResponse

Message for response to listing Instances

ListSupportedDatabaseFlagsRequest

Message for listing the information about the supported Database flags.

ListSupportedDatabaseFlagsResponse

Message for response to listing SupportedDatabaseFlags.

ListUsersRequest

Message for requesting list of Users

ListUsersResponse

Message for response to listing Users

MigrationSource

Subset of the source instance configuration that is available when reading the cluster resource.

OperationMetadata

Represents the metadata of the long-running operation.

PromoteClusterRequest

Message for promoting a Cluster

RestartInstanceRequest

RestoreClusterRequest

Message for restoring a Cluster from a backup or another cluster at a given point in time.

SslConfig

SSL configuration.

SupportedDatabaseFlag

SupportedDatabaseFlag gives general information about a database flag, like type and allowed values. This is a static value that is defined on the server side, and it cannot be modified by callers. To set the Database flags on a particular Instance, a caller should modify the Instance.database_flags field.

UpdateBackupRequest

Message for updating a Backup

UpdateClusterRequest

Message for updating a Cluster

UpdateInstanceRequest

Message for updating a Instance

UpdateUserRequest

Message for updating a User

User

Message describing User object.

UserPassword

The username/password for a database user. Used for specifying initial users at cluster creation time.

Enums

ClusterView

View on Cluster. Pass this enum to rpcs that returns a cluster message to control which subsets of fields to get.

DatabaseVersion

The supported database engine versions.

InstanceView

View on Instance. Pass this enum to rpcs that returns an Instance message to control which subsets of fields to get.