Module google_api_proto::google::cloud::apigateway::v1

Modules

api

Nested message and enum types in Api.

api_config

Nested message and enum types in ApiConfig.

api_gateway_service_client

Generated client implementations.

gateway

Nested message and enum types in Gateway.

get_api_config_request

Nested message and enum types in GetApiConfigRequest.

operation_metadata

Nested message and enum types in OperationMetadata.

Structs

Api

An API that can be served by one or more Gateways.

ApiConfig

An API Configuration is a combination of settings for both the Managed Service and Gateways serving this API Config.

CreateApiConfigRequest

Request message for ApiGatewayService.CreateApiConfig

CreateApiRequest

Request message for ApiGatewayService.CreateApi

CreateGatewayRequest

Request message for ApiGatewayService.CreateGateway

DeleteApiConfigRequest

Request message for ApiGatewayService.DeleteApiConfig

DeleteApiRequest

Request message for ApiGatewayService.DeleteApi

DeleteGatewayRequest

Request message for ApiGatewayService.DeleteGateway

Gateway

A Gateway is an API-aware HTTP proxy. It performs API-Method and/or API-Consumer specific actions based on an API Config such as authentication, policy enforcement, and backend selection.

GetApiConfigRequest

Request message for ApiGatewayService.GetApiConfig

GetApiRequest

Request message for ApiGatewayService.GetApi

GetGatewayRequest

Request message for ApiGatewayService.GetGateway

ListApiConfigsRequest

Request message for ApiGatewayService.ListApiConfigs

ListApiConfigsResponse

Response message for ApiGatewayService.ListApiConfigs

ListApisRequest

Request message for ApiGatewayService.ListApis

ListApisResponse

Response message for ApiGatewayService.ListApis

ListGatewaysRequest

Request message for ApiGatewayService.ListGateways

ListGatewaysResponse

Response message for ApiGatewayService.ListGateways

OperationMetadata

Represents the metadata of the long-running operation.

UpdateApiConfigRequest

Request message for ApiGatewayService.UpdateApiConfig

UpdateApiRequest

Request message for ApiGatewayService.UpdateApi

UpdateGatewayRequest

Request message for ApiGatewayService.UpdateGateway