Struct google_api_proto::google::cloud::apigateway::v1::api_gateway_service_client::ApiGatewayServiceClient

pub struct ApiGatewayServiceClient<T> { /* private fields */ }

The API Gateway Service is the interface for managing API Gateways.

Implementations

impl<T> ApiGatewayServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> ApiGatewayServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_gateways( &mut self, request: impl IntoRequest<ListGatewaysRequest> ) -> Result<Response<ListGatewaysResponse>, Status>

Lists Gateways in a given project and location.

pub async fn get_gateway( &mut self, request: impl IntoRequest<GetGatewayRequest> ) -> Result<Response<Gateway>, Status>

Gets details of a single Gateway.

pub async fn create_gateway( &mut self, request: impl IntoRequest<CreateGatewayRequest> ) -> Result<Response<Operation>, Status>

Creates a new Gateway in a given project and location.

pub async fn update_gateway( &mut self, request: impl IntoRequest<UpdateGatewayRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Gateway.

pub async fn delete_gateway( &mut self, request: impl IntoRequest<DeleteGatewayRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Gateway.

pub async fn list_apis( &mut self, request: impl IntoRequest<ListApisRequest> ) -> Result<Response<ListApisResponse>, Status>

Lists Apis in a given project and location.

pub async fn get_api( &mut self, request: impl IntoRequest<GetApiRequest> ) -> Result<Response<Api>, Status>

Gets details of a single Api.

pub async fn create_api( &mut self, request: impl IntoRequest<CreateApiRequest> ) -> Result<Response<Operation>, Status>

Creates a new Api in a given project and location.

pub async fn update_api( &mut self, request: impl IntoRequest<UpdateApiRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Api.

pub async fn delete_api( &mut self, request: impl IntoRequest<DeleteApiRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Api.

pub async fn list_api_configs( &mut self, request: impl IntoRequest<ListApiConfigsRequest> ) -> Result<Response<ListApiConfigsResponse>, Status>

Lists ApiConfigs in a given project and location.

pub async fn get_api_config( &mut self, request: impl IntoRequest<GetApiConfigRequest> ) -> Result<Response<ApiConfig>, Status>

Gets details of a single ApiConfig.

pub async fn create_api_config( &mut self, request: impl IntoRequest<CreateApiConfigRequest> ) -> Result<Response<Operation>, Status>

Creates a new ApiConfig in a given project and location.

pub async fn update_api_config( &mut self, request: impl IntoRequest<UpdateApiConfigRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single ApiConfig.

pub async fn delete_api_config( &mut self, request: impl IntoRequest<DeleteApiConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes a single ApiConfig.