Module google_api_proto::google::cloud::apihub::v1

Modules

api_hub_client

Generated client implementations.

api_hub_dependencies_client

Generated client implementations.

api_hub_instance

Nested message and enum types in ApiHubInstance.

api_hub_plugin_client

Generated client implementations.

api_hub_resource

Nested message and enum types in ApiHubResource.

attribute

Nested message and enum types in Attribute.

attribute_values

Nested message and enum types in AttributeValues.

definition

Nested message and enum types in Definition.

dependency

Nested message and enum types in Dependency.

dependency_entity_reference

Nested message and enum types in DependencyEntityReference.

dependency_error_detail

Nested message and enum types in DependencyErrorDetail.

host_project_registration_service_client

Generated client implementations.

http_operation

Nested message and enum types in HttpOperation.

lint_response

Nested message and enum types in LintResponse.

linting_service_client

Generated client implementations.

open_api_spec_details

Nested message and enum types in OpenApiSpecDetails.

operation_details

Nested message and enum types in OperationDetails.

plugin

Nested message and enum types in Plugin.

provisioning_client

Generated client implementations.

runtime_project_attachment_service_client

Generated client implementations.

spec

Nested message and enum types in Spec.

spec_details

Nested message and enum types in SpecDetails.

Structs

Api

An API resource in the API Hub.

ApiHubInstance

An ApiHubInstance represents the instance resources of the API Hub. Currently, only one ApiHub instance is allowed for each project.

ApiHubResource

ApiHubResource is one of the resources such as Api, Operation, Deployment, Definition, Spec and Version resources stored in API-Hub.

ApiOperation

Represents an operation contained in an API version in the API Hub. An operation is added/updated/deleted in an API version when a new spec is added or an existing spec is updated/deleted in a version. Currently, an operation will be created only corresponding to OpenAPI spec as parsing is supported for OpenAPI spec.

Attribute

An attribute in the API Hub. An attribute is a name value pair which can be attached to different resources in the API hub based on the scope of the attribute. Attributes can either be pre-defined by the API Hub or created by users.

AttributeValues

The attribute values associated with resource.

CreateApiHubInstanceRequest

The [CreateApiHubInstance][google.cloud.apihub.v1.Provisioning.CreateApiHubInstance] method’s request.

CreateApiRequest

The [CreateApi][google.cloud.apihub.v1.ApiHub.CreateApi] method’s request.

CreateAttributeRequest

The [CreateAttribute][google.cloud.apihub.v1.ApiHub.CreateAttribute] method’s request.

CreateDependencyRequest

The [CreateDependency][google.cloud.apihub.v1.ApiHubDependencies.CreateDependency] method’s request.

CreateDeploymentRequest

The [CreateDeployment][google.cloud.apihub.v1.ApiHub.CreateDeployment] method’s request.

CreateExternalApiRequest

The [CreateExternalApi][google.cloud.apihub.v1.ApiHub.CreateExternalApi] method’s request.

CreateHostProjectRegistrationRequest

The [CreateHostProjectRegistration][google.cloud.apihub.v1.HostProjectRegistrationService.CreateHostProjectRegistration] method’s request.

CreateRuntimeProjectAttachmentRequest

The [CreateRuntimeProjectAttachment][google.cloud.apihub.v1.RuntimeProjectAttachmentService.CreateRuntimeProjectAttachment] method’s request.

CreateSpecRequest

The [CreateSpec][google.cloud.apihub.v1.ApiHub.CreateSpec] method’s request.

CreateVersionRequest

The [CreateVersion][google.cloud.apihub.v1.ApiHub.CreateVersion] method’s request.

Definition

Represents a definition for example schema, request, response definitions contained in an API version. A definition is added/updated/deleted in an API version when a new spec is added or an existing spec is updated/deleted in a version. Currently, definition will be created only corresponding to OpenAPI spec as parsing is supported for OpenAPI spec. Also, within OpenAPI spec, only schema object is supported.

DeleteApiRequest

The [DeleteApi][google.cloud.apihub.v1.ApiHub.DeleteApi] method’s request.

DeleteAttributeRequest

The [DeleteAttribute][google.cloud.apihub.v1.ApiHub.DeleteAttribute] method’s request.

DeleteDependencyRequest

The [DeleteDependency][google.cloud.apihub.v1.ApiHubDependencies.DeleteDependency] method’s request.

DeleteDeploymentRequest

The [DeleteDeployment][google.cloud.apihub.v1.ApiHub.DeleteDeployment] method’s request.

DeleteExternalApiRequest

The [DeleteExternalApi][google.cloud.apihub.v1.ApiHub.DeleteExternalApi] method’s request.

DeleteRuntimeProjectAttachmentRequest

The [DeleteRuntimeProjectAttachment][google.cloud.apihub.v1.RuntimeProjectAttachmentService.DeleteRuntimeProjectAttachment] method’s request.

DeleteSpecRequest

The [DeleteSpec][google.cloud.apihub.v1.ApiHub.DeleteSpec] method’s request.

DeleteVersionRequest

The [DeleteVersion][google.cloud.apihub.v1.ApiHub.DeleteVersion] method’s request.

Dependency

A dependency resource defined in the API hub describes a dependency directed from a consumer to a supplier entity. A dependency can be defined between two [Operations][google.cloud.apihub.v1.Operation] or between an [Operation][google.cloud.apihub.v1.Operation] and [External API][google.cloud.apihub.v1.ExternalApi].

DependencyEntityReference

Reference to an entity participating in a dependency.

DependencyErrorDetail

Details describing error condition of a dependency.

Deployment

Details of the deployment where APIs are hosted. A deployment could represent an Apigee proxy, API gateway, other Google Cloud services or non-Google Cloud services as well. A deployment entity is a root level entity in the API hub and exists independent of any API.

DisablePluginRequest

The [DisablePlugin][google.cloud.apihub.v1.ApiHubPlugin.DisablePlugin] method’s request.

Documentation

Documentation details.

EnablePluginRequest

The [EnablePlugin][google.cloud.apihub.v1.ApiHubPlugin.EnablePlugin] method’s request.

ExternalApi

An external API represents an API being provided by external sources. This can be used to model third-party APIs and can be used to define dependencies.

GetApiHubInstanceRequest

The [GetApiHubInstance][google.cloud.apihub.v1.Provisioning.GetApiHubInstance] method’s request.

GetApiOperationRequest

The [GetApiOperation][google.cloud.apihub.v1.ApiHub.GetApiOperation] method’s request.

GetApiRequest

The [GetApi][google.cloud.apihub.v1.ApiHub.GetApi] method’s request.

GetAttributeRequest

The [GetAttribute][google.cloud.apihub.v1.ApiHub.GetAttribute] method’s request.

GetDefinitionRequest

The [GetDefinition][google.cloud.apihub.v1.ApiHub.GetDefinition] method’s request.

GetDependencyRequest

The [GetDependency][.ApiHubDependencies.GetDependency] method’s request.

GetDeploymentRequest

The [GetDeployment][google.cloud.apihub.v1.ApiHub.GetDeployment] method’s request.

GetExternalApiRequest

The [GetExternalApi][google.cloud.apihub.v1.ApiHub.GetExternalApi] method’s request.

GetHostProjectRegistrationRequest

The [GetHostProjectRegistration][google.cloud.apihub.v1.HostProjectRegistrationService.GetHostProjectRegistration] method’s request.

GetPluginRequest

The [GetPlugin][google.cloud.apihub.v1.ApiHubPlugin.GetPlugin] method’s request.

GetRuntimeProjectAttachmentRequest

The [GetRuntimeProjectAttachment][google.cloud.apihub.v1.RuntimeProjectAttachmentService.GetRuntimeProjectAttachment] method’s request.

GetSpecContentsRequest

The [GetSpecContents][google.cloud.apihub.v1.ApiHub.GetSpecContents] method’s request.

GetSpecRequest

The [GetSpec][google.cloud.apihub.v1.ApiHub.GetSpec] method’s request.

GetStyleGuideContentsRequest

The [GetStyleGuideContents][ApiHub.GetStyleGuideContents] method’s request.

GetStyleGuideRequest

The [GetStyleGuide][ApiHub.GetStyleGuide] method’s request.

GetVersionRequest

The [GetVersion][google.cloud.apihub.v1.ApiHub.GetVersion] method’s request.

HostProjectRegistration

Host project registration refers to the registration of a Google cloud project with Api Hub as a host project. This is the project where Api Hub is provisioned. It acts as the consumer project for the Api Hub instance provisioned. Multiple runtime projects can be attached to the host project and these attachments define the scope of Api Hub.

HttpOperation

The HTTP Operation.

Issue

Issue contains the details of a single issue found by the linter.

LintResponse

LintResponse contains the response from the linter.

LintSpecRequest

The [LintSpec][ApiHub.LintSpec] method’s request.

ListApiOperationsRequest

The [ListApiOperations][google.cloud.apihub.v1.ApiHub.ListApiOperations] method’s request.

ListApiOperationsResponse

The [ListApiOperations][google.cloud.apihub.v1.ApiHub.ListApiOperations] method’s response.

ListApisRequest

The [ListApis][google.cloud.apihub.v1.ApiHub.ListApis] method’s request.

ListApisResponse

The [ListApis][google.cloud.apihub.v1.ApiHub.ListApis] method’s response.

ListAttributesRequest

The [ListAttributes][google.cloud.apihub.v1.ApiHub.ListAttributes] method’s request.

ListAttributesResponse

The [ListAttributes][google.cloud.apihub.v1.ApiHub.ListAttributes] method’s response.

ListDependenciesRequest

The [ListDependencies][google.cloud.apihub.v1.ApiHubDependencies.ListDependencies] method’s request.

ListDependenciesResponse

The [ListDependencies][google.cloud.apihub.v1.ApiHubDependencies.ListDependencies] method’s response.

ListDeploymentsRequest

The [ListDeployments][google.cloud.apihub.v1.ApiHub.ListDeployments] method’s request.

ListDeploymentsResponse

The [ListDeployments][google.cloud.apihub.v1.ApiHub.ListDeployments] method’s response.

ListExternalApisRequest

The [ListExternalApis][google.cloud.apihub.v1.ApiHub.ListExternalApis] method’s request.

ListExternalApisResponse

The [ListExternalApis][google.cloud.apihub.v1.ApiHub.ListExternalApis] method’s response.

ListHostProjectRegistrationsRequest

The [ListHostProjectRegistrations][google.cloud.apihub.v1.HostProjectRegistrationService.ListHostProjectRegistrations] method’s request.

ListHostProjectRegistrationsResponse

The [ListHostProjectRegistrations][google.cloud.apihub.v1.HostProjectRegistrationService.ListHostProjectRegistrations] method’s response.

ListRuntimeProjectAttachmentsRequest

The [ListRuntimeProjectAttachments][google.cloud.apihub.v1.RuntimeProjectAttachmentService.ListRuntimeProjectAttachments] method’s request.

ListRuntimeProjectAttachmentsResponse

The [ListRuntimeProjectAttachments][google.cloud.apihub.v1.RuntimeProjectAttachmentService.ListRuntimeProjectAttachments] method’s response.

ListSpecsRequest

The [ListSpecs][ListSpecs] method’s request.

ListSpecsResponse

The [ListSpecs][google.cloud.apihub.v1.ApiHub.ListSpecs] method’s response.

ListVersionsRequest

The [ListVersions][google.cloud.apihub.v1.ApiHub.ListVersions] method’s request.

ListVersionsResponse

The [ListVersions][google.cloud.apihub.v1.ApiHub.ListVersions] method’s response.

LookupApiHubInstanceRequest

The [LookupApiHubInstance][google.cloud.apihub.v1.Provisioning.LookupApiHubInstance] method’s request.

LookupApiHubInstanceResponse

The [LookupApiHubInstance][google.cloud.apihub.v1.Provisioning.LookupApiHubInstance] method’s response.`

LookupRuntimeProjectAttachmentRequest

The [LookupRuntimeProjectAttachment][google.cloud.apihub.v1.RuntimeProjectAttachmentService.LookupRuntimeProjectAttachment] method’s request.

LookupRuntimeProjectAttachmentResponse

The [ListRuntimeProjectAttachments][google.cloud.apihub.v1.RuntimeProjectAttachmentService.ListRuntimeProjectAttachments] method’s response.

OpenApiSpecDetails

OpenApiSpecDetails contains the details parsed from an OpenAPI spec in addition to the fields mentioned in [SpecDetails][google.cloud.apihub.v1.SpecDetails].

OperationDetails

The operation details parsed from the spec.

OperationMetadata

Represents the metadata of the long-running operation.

Owner

Owner details.

Path

The path details derived from the spec.

Plugin

A plugin resource in the API Hub.

Point

Point within the file (line and character).

Range

Object describing where in the file the issue was found.

RuntimeProjectAttachment

Runtime project attachment represents an attachment from the runtime project to the host project. Api Hub looks for deployments in the attached runtime projects and creates corresponding resources in Api Hub for the discovered deployments.

Schema

The schema details derived from the spec. Currently, this entity is supported for OpenAPI spec only. For OpenAPI spec, this maps to the schema defined in the definitions section for OpenAPI 2.0 version and in components.schemas section for OpenAPI 3.0 and 3.1 version.

SearchResourcesRequest

The [SearchResources][google.cloud.apihub.v1.ApiHub.SearchResources] method’s request.

SearchResourcesResponse

Response for the [SearchResources][google.cloud.apihub.v1.ApiHub.SearchResources] method.

SearchResult

Represents the search results.

Spec

Represents a spec associated with an API version in the API Hub. Note that specs of various types can be uploaded, however parsing of details is supported for OpenAPI spec currently.

SpecContents

The spec contents.

SpecDetails

SpecDetails contains the details parsed from supported spec types.

StyleGuide

Represents a singleton style guide resource to be used for linting Open API specs.

StyleGuideContents

The style guide contents.

UpdateApiRequest

The [UpdateApi][google.cloud.apihub.v1.ApiHub.UpdateApi] method’s request.

UpdateAttributeRequest

The [UpdateAttribute][google.cloud.apihub.v1.ApiHub.UpdateAttribute] method’s request.

UpdateDependencyRequest

The [UpdateDependency][google.cloud.apihub.v1.ApiHubDependencies.UpdateDependency] method’s request.

UpdateDeploymentRequest

The [UpdateDeployment][google.cloud.apihub.v1.ApiHub.UpdateDeployment] method’s request.

UpdateExternalApiRequest

The [UpdateExternalApi][google.cloud.apihub.v1.ApiHub.UpdateExternalApi] method’s request.

UpdateSpecRequest

The [UpdateSpec][google.cloud.apihub.v1.ApiHub.UpdateSpec] method’s request.

UpdateStyleGuideRequest

The [UpdateStyleGuide][ApiHub.UpdateStyleGuide] method’s request.

UpdateVersionRequest

The [UpdateVersion][google.cloud.apihub.v1.ApiHub.UpdateVersion] method’s request.

Version

Represents a version of the API resource in API hub. This is also referred to as the API version.

Enums

LintState

Lint state represents success or failure for linting.

Linter

Enumeration of linter types.

Severity

Severity of the issue.