Module google_api_proto::google::cloud::apphub::v1

Modules

app_hub_client

Generated client implementations.

application

Nested message and enum types in Application.

criticality

Nested message and enum types in Criticality.

environment

Nested message and enum types in Environment.

scope

Nested message and enum types in Scope.

service

Nested message and enum types in Service.

service_project_attachment

Nested message and enum types in ServiceProjectAttachment.

workload

Nested message and enum types in Workload.

Structs

Application

Application defines the governance boundary for App Hub Entities that perform a logical end-to-end business function. App Hub supports application level IAM permission to align with governance requirements.

Attributes

Consumer provided attributes.

ContactInfo

Contact information of stakeholders.

CreateApplicationRequest

Request for CreateApplication.

CreateServiceProjectAttachmentRequest

Request for CreateServiceProjectAttachment.

CreateServiceRequest

Request for CreateService.

CreateWorkloadRequest

Request for CreateWorkload.

Criticality

Criticality of the Application, Service, or Workload

DeleteApplicationRequest

Request for DeleteApplication.

DeleteServiceProjectAttachmentRequest

Request for DeleteServiceProjectAttachment.

DeleteServiceRequest

Request for DeleteService.

DeleteWorkloadRequest

Request for DeleteWorkload.

DetachServiceProjectAttachmentRequest

Request for DetachServiceProjectAttachment.

DetachServiceProjectAttachmentResponse

Response for DetachServiceProjectAttachment.

DiscoveredService

DiscoveredService is a network/api interface that exposes some functionality to clients for consumption over the network. A discovered service can be registered to a App Hub service.

DiscoveredWorkload

DiscoveredWorkload is a binary deployment (such as managed instance groups (MIGs) and GKE deployments) that performs the smallest logical subset of business functionality. A discovered workload can be registered to an App Hub Workload.

Environment

Environment of the Application, Service, or Workload

GetApplicationRequest

Request for GetApplication.

GetDiscoveredServiceRequest

Request for GetDiscoveredService.

GetDiscoveredWorkloadRequest

Request for GetDiscoveredWorkload.

GetServiceProjectAttachmentRequest

Request for GetServiceProjectAttachment.

GetServiceRequest

Request for GetService.

GetWorkloadRequest

Request for GetWorkload.

ListApplicationsRequest

Request for ListApplications.

ListApplicationsResponse

Response for ListApplications.

ListDiscoveredServicesRequest

Request for ListDiscoveredServices.

ListDiscoveredServicesResponse

Response for ListDiscoveredServices.

ListDiscoveredWorkloadsRequest

Request for ListDiscoveredWorkloads.

ListDiscoveredWorkloadsResponse

Response for ListDiscoveredWorkloads.

ListServiceProjectAttachmentsRequest

Request for ListServiceProjectAttachments.

ListServiceProjectAttachmentsResponse

Response for ListServiceProjectAttachments.

ListServicesRequest

Request for ListServices.

ListServicesResponse

Response for ListServices.

ListWorkloadsRequest

Request for ListWorkloads.

ListWorkloadsResponse

Response for ListWorkloads.

LookupDiscoveredServiceRequest

Request for LookupDiscoveredService.

LookupDiscoveredServiceResponse

Response for LookupDiscoveredService.

LookupDiscoveredWorkloadRequest

Request for LookupDiscoveredWorkload.

LookupDiscoveredWorkloadResponse

Response for LookupDiscoveredWorkload.

LookupServiceProjectAttachmentRequest

Request for LookupServiceProjectAttachment.

LookupServiceProjectAttachmentResponse

Response for LookupServiceProjectAttachment.

OperationMetadata

Represents the metadata of the long-running operation.

Scope

Scope of an application.

Service

Service is an App Hub data model that contains a discovered service, which represents a network/api interface that exposes some functionality to clients for consumption over the network.

ServiceProjectAttachment

ServiceProjectAttachment represents an attachment from a service project to a host project. Service projects contain the underlying cloud infrastructure resources, and expose these resources to the host project through a ServiceProjectAttachment. With the attachments, the host project can provide an aggregated view of resources across all service projects.

ServiceProperties

Properties of an underlying cloud resource that can comprise a Service.

ServiceReference

Reference to an underlying networking resource that can comprise a Service.

UpdateApplicationRequest

Request for UpdateApplication.

UpdateServiceRequest

Request for UpdateService.

UpdateWorkloadRequest

Request for UpdateWorkload.

Workload

Workload is an App Hub data model that contains a discovered workload, which represents a binary deployment (such as managed instance groups (MIGs) and GKE deployments) that performs the smallest logical subset of business functionality.

WorkloadProperties

Properties of an underlying compute resource represented by the Workload.

WorkloadReference

Reference of an underlying compute resource represented by the Workload.