Module google_api_proto::google::cloud::asset::v1

Modules

analyze_iam_policy_response

Nested message and enum types in AnalyzeIamPolicyResponse.

analyze_move_request

Nested message and enum types in AnalyzeMoveRequest.

analyze_org_policies_response

Nested message and enum types in AnalyzeOrgPoliciesResponse.

analyze_org_policy_governed_assets_response

Nested message and enum types in AnalyzeOrgPolicyGovernedAssetsResponse.

analyze_org_policy_governed_containers_response

Nested message and enum types in AnalyzeOrgPolicyGovernedContainersResponse.

analyzer_org_policy

Nested message and enum types in AnalyzerOrgPolicy.

analyzer_org_policy_constraint

Nested message and enum types in AnalyzerOrgPolicyConstraint.

asset

Nested message and enum types in Asset.

asset_service_client

Generated client implementations.

batch_get_effective_iam_policies_response

Nested message and enum types in BatchGetEffectiveIamPoliciesResponse.

condition_evaluation

Nested message and enum types in ConditionEvaluation.

feed_output_config

Nested message and enum types in FeedOutputConfig.

gcs_destination

Nested message and enum types in GcsDestination.

iam_policy_analysis_output_config

Nested message and enum types in IamPolicyAnalysisOutputConfig.

iam_policy_analysis_query

Nested message and enum types in IamPolicyAnalysisQuery.

iam_policy_analysis_result

Nested message and enum types in IamPolicyAnalysisResult.

iam_policy_search_result

Nested message and enum types in IamPolicySearchResult.

move_analysis

Nested message and enum types in MoveAnalysis.

output_config

Nested message and enum types in OutputConfig.

output_result

Nested message and enum types in OutputResult.

partition_spec

Nested message and enum types in PartitionSpec.

query_assets_output_config

Nested message and enum types in QueryAssetsOutputConfig.

query_assets_request

Nested message and enum types in QueryAssetsRequest.

query_assets_response

Nested message and enum types in QueryAssetsResponse.

saved_query

Nested message and enum types in SavedQuery.

temporal_asset

Nested message and enum types in TemporalAsset.

Structs

AnalyzeIamPolicyLongrunningMetadata

Represents the metadata of the longrunning operation for the AnalyzeIamPolicyLongrunning RPC.

AnalyzeIamPolicyLongrunningRequest

A request message for [AssetService.AnalyzeIamPolicyLongrunning][google.cloud.asset.v1.AssetService.AnalyzeIamPolicyLongrunning].

AnalyzeIamPolicyLongrunningResponse

A response message for [AssetService.AnalyzeIamPolicyLongrunning][google.cloud.asset.v1.AssetService.AnalyzeIamPolicyLongrunning].

AnalyzeIamPolicyRequest

A request message for [AssetService.AnalyzeIamPolicy][google.cloud.asset.v1.AssetService.AnalyzeIamPolicy].

AnalyzeIamPolicyResponse

A response message for [AssetService.AnalyzeIamPolicy][google.cloud.asset.v1.AssetService.AnalyzeIamPolicy].

AnalyzeMoveRequest

The request message for performing resource move analysis.

AnalyzeMoveResponse

The response message for resource move analysis.

AnalyzeOrgPoliciesRequest

A request message for [AssetService.AnalyzeOrgPolicies][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicies].

AnalyzeOrgPoliciesResponse

The response message for [AssetService.AnalyzeOrgPolicies][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicies].

AnalyzeOrgPolicyGovernedAssetsRequest

A request message for [AssetService.AnalyzeOrgPolicyGovernedAssets][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicyGovernedAssets].

AnalyzeOrgPolicyGovernedAssetsResponse

The response message for [AssetService.AnalyzeOrgPolicyGovernedAssets][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicyGovernedAssets].

AnalyzeOrgPolicyGovernedContainersRequest

A request message for [AssetService.AnalyzeOrgPolicyGovernedContainers][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicyGovernedContainers].

AnalyzeOrgPolicyGovernedContainersResponse

The response message for [AssetService.AnalyzeOrgPolicyGovernedContainers][google.cloud.asset.v1.AssetService.AnalyzeOrgPolicyGovernedContainers].

AnalyzerOrgPolicy

This organization policy message is a modified version of the one defined in the Organization Policy system. This message contains several fields defined in the original organization policy with some new fields for analysis purpose.

AnalyzerOrgPolicyConstraint

The organization policy constraint definition.

Asset

An asset in Google Cloud. An asset can be any resource in the Google Cloud resource hierarchy, a resource outside the Google Cloud resource hierarchy (such as Google Kubernetes Engine clusters and objects), or a policy (e.g. IAM policy), or a relationship (e.g. an INSTANCE_TO_INSTANCEGROUP relationship). See Supported asset types for more information.

AttachedResource

Attached resource representation, which is defined by the corresponding service provider. It represents an attached resource’s payload.

BatchGetAssetsHistoryRequest

Batch get assets history request.

BatchGetAssetsHistoryResponse

Batch get assets history response.

BatchGetEffectiveIamPoliciesRequest

A request message for [AssetService.BatchGetEffectiveIamPolicies][google.cloud.asset.v1.AssetService.BatchGetEffectiveIamPolicies].

BatchGetEffectiveIamPoliciesResponse

A response message for [AssetService.BatchGetEffectiveIamPolicies][google.cloud.asset.v1.AssetService.BatchGetEffectiveIamPolicies].

BigQueryDestination

A BigQuery destination for exporting assets to.

ConditionEvaluation

The condition evaluation.

CreateFeedRequest

Create asset feed request.

CreateSavedQueryRequest

Request to create a saved query.

DeleteFeedRequest

DeleteSavedQueryRequest

Request to delete a saved query.

EffectiveTagDetails

The effective tags and the ancestor resources from which they were inherited.

ExportAssetsRequest

Export asset request.

ExportAssetsResponse

The export asset response. This message is returned by the [google.longrunning.Operations.GetOperation][google.longrunning.Operations.GetOperation] method in the returned [google.longrunning.Operation.response][google.longrunning.Operation.response] field.

Feed

An asset feed used to export asset updates to a destinations. An asset feed filter controls what updates are exported. The asset feed must be created within a project, organization, or folder. Supported destinations are: Pub/Sub topics.

FeedOutputConfig

Output configuration for asset feed destination.

GcsDestination

A Cloud Storage location.

GcsOutputResult

A Cloud Storage output result.

GetFeedRequest

Get asset feed request.

GetSavedQueryRequest

Request to get a saved query.

IamPolicyAnalysisOutputConfig

Output configuration for export IAM policy analysis destination.

IamPolicyAnalysisQuery

IAM policy analysis query message.

IamPolicyAnalysisResult

IAM Policy analysis result, consisting of one IAM policy binding and derived access control lists.

IamPolicyAnalysisState

Represents the detailed state of an entity under analysis, such as a resource, an identity or an access.

IamPolicySearchResult

A result of IAM Policy search, containing information of an IAM policy.

ListAssetsRequest

ListAssets request.

ListAssetsResponse

ListAssets response.

ListFeedsRequest

List asset feeds request.

ListFeedsResponse

ListSavedQueriesRequest

Request to list saved queries.

ListSavedQueriesResponse

Response of listing saved queries.

MoveAnalysis

A message to group the analysis information.

MoveAnalysisResult

An analysis result including blockers and warnings.

MoveImpact

A message to group impacts of moving the target resource.

OutputConfig

Output configuration for export assets destination.

OutputResult

Output result of export assets.

PartitionSpec

Specifications of BigQuery partitioned table as export destination.

PubsubDestination

A Pub/Sub destination.

QueryAssetsOutputConfig

Output configuration query assets.

QueryAssetsRequest

QueryAssets request.

QueryAssetsResponse

QueryAssets response.

QueryResult

Execution results of the query.

RelatedAsset

An asset identifier in Google Cloud which contains its name, type and ancestors. An asset can be any resource in the Google Cloud resource hierarchy, a resource outside the Google Cloud resource hierarchy (such as Google Kubernetes Engine clusters and objects), or a policy (e.g. IAM policy). See Supported asset types for more information.

RelatedAssets

DEPRECATED. This message only presents for the purpose of backward-compatibility. The server will never populate this message in responses. The detailed related assets with the relationship_type.

RelatedResource

The detailed related resource.

RelatedResources

The related resources of the primary resource.

RelationshipAttributes

DEPRECATED. This message only presents for the purpose of backward-compatibility. The server will never populate this message in responses. The relationship attributes which include type, source_resource_type, target_resource_type and action.

Resource

A representation of a Google Cloud resource.

ResourceSearchResult

A result of Resource Search, containing information of a cloud resource. Next ID: 34

SavedQuery

A saved query which can be shared with others or used later.

SearchAllIamPoliciesRequest

Search all IAM policies request.

SearchAllIamPoliciesResponse

Search all IAM policies response.

SearchAllResourcesRequest

Search all resources request.

SearchAllResourcesResponse

Search all resources response.

TableFieldSchema

A field in TableSchema.

TableSchema

BigQuery Compatible table schema.

Tag

The key and value for a tag.

TemporalAsset

An asset in Google Cloud and its temporal metadata, including the time window when it was observed and its status during that window.

TimeWindow

A time window specified by its start_time and end_time.

UpdateFeedRequest

Update asset feed request.

UpdateSavedQueryRequest

Request to update a saved query.

VersionedResource

Resource representation as defined by the corresponding service providing the resource for a given API version.

Enums

ContentType

Asset content type.