Module google_api_proto::google::cloud::asset::v1p2beta1

Modules

asset

Nested message and enum types in Asset.

asset_service_client

Generated client implementations.

feed_output_config

Nested message and enum types in FeedOutputConfig.

gcs_destination

Nested message and enum types in GcsDestination.

output_config

Nested message and enum types in OutputConfig.

Structs

Asset

An asset in Google Cloud. An asset can be any resource in the Google Cloud resource hierarchy, a resource outside the Google Cloud resource hierarchy (such as Google Kubernetes Engine clusters and objects), or a policy (e.g. IAM policy). See Supported asset types for more information.

BatchGetAssetsHistoryResponse

Batch get assets history response.

CreateFeedRequest

Create asset feed request.

DeleteFeedRequest

ExportAssetsResponse

The export asset response. This message is returned by the [google.longrunning.Operations.GetOperation][google.longrunning.Operations.GetOperation] method in the returned [google.longrunning.Operation.response][google.longrunning.Operation.response] field.

Feed

An asset feed used to export asset updates to a destinations. An asset feed filter controls what updates are exported. The asset feed must be created within a project, organization, or folder. Supported destinations are: Cloud Pub/Sub topics.

FeedOutputConfig

Output configuration for asset feed destination.

GcsDestination

A Cloud Storage location.

GetFeedRequest

Get asset feed request.

ListFeedsRequest

List asset feeds request.

ListFeedsResponse

OutputConfig

Output configuration for export assets destination.

PubsubDestination

A Pub/Sub destination.

Resource

A representation of a Google Cloud resource.

TemporalAsset

An asset in Google Cloud and its temporal metadata, including the time window when it was observed and its status during that window.

TimeWindow

A time window specified by its start_time and end_time.

UpdateFeedRequest

Update asset feed request.

Enums

ContentType

Asset content type.