Module google_api_proto::google::cloud::assuredworkloads::v1

Modules

assured_workloads_service_client

Generated client implementations.

restrict_allowed_resources_request

Nested message and enum types in RestrictAllowedResourcesRequest.

violation

Nested message and enum types in Violation.

workload

Nested message and enum types in Workload.

Structs

AcknowledgeViolationRequest

Request for acknowledging the violation Next Id: 4

AcknowledgeViolationResponse

Response for violation acknowledgement

CreateWorkloadOperationMetadata

Operation metadata to give request details of CreateWorkload.

CreateWorkloadRequest

Request for creating a workload.

DeleteWorkloadRequest

Request for deleting a Workload.

GetViolationRequest

Request for fetching a Workload Violation.

GetWorkloadRequest

Request for fetching a workload.

ListViolationsRequest

Request for fetching violations in an organization.

ListViolationsResponse

Response of ListViolations endpoint.

ListWorkloadsRequest

Request for fetching workloads in an organization.

ListWorkloadsResponse

Response of ListWorkloads endpoint.

RestrictAllowedResourcesRequest

Request for restricting list of available resources in Workload environment.

RestrictAllowedResourcesResponse

Response for restricting the list of allowed resources.

TimeWindow

Interval defining a time window.

UpdateWorkloadRequest

Request for Updating a workload.

Violation

Workload monitoring Violation.

Workload

A Workload object for managing highly regulated workloads of cloud customers.