Module google_api_proto::google::cloud::assuredworkloads::v1beta1

Modules

analyze_workload_move_request

Nested message and enum types in AnalyzeWorkloadMoveRequest.

assured_workloads_service_client

Generated client implementations.

restrict_allowed_resources_request

Nested message and enum types in RestrictAllowedResourcesRequest.

workload

Nested message and enum types in Workload.

Structs

AnalyzeWorkloadMoveRequest

A request to analyze a hypothetical move of a source project or project-based workload to a target (destination) folder-based workload.

AnalyzeWorkloadMoveResponse

A response that includes the analysis of the hypothetical resource move.

CreateWorkloadOperationMetadata

Operation metadata to give request details of CreateWorkload.

CreateWorkloadRequest

Request for creating a workload.

DeleteWorkloadRequest

Request for deleting a Workload.

GetWorkloadRequest

Request for fetching a workload.

ListWorkloadsRequest

Request for fetching workloads in an organization.

ListWorkloadsResponse

Response of ListWorkloads endpoint.

RestrictAllowedResourcesRequest

Request for restricting list of available resources in Workload environment.

RestrictAllowedResourcesResponse

Response for restricting the list of allowed resources.

UpdateWorkloadRequest

Request for Updating a workload.

Workload

An Workload object for managing highly regulated workloads of cloud customers.