Module google_api_proto::google::cloud::audit

Modules

big_query_audit_metadata

Nested message and enum types in BigQueryAuditMetadata.

service_account_delegation_info

Nested message and enum types in ServiceAccountDelegationInfo.

violation_info

Nested message and enum types in ViolationInfo.

Structs

AuditLog

Common audit log format for Google Cloud Platform API operations.

AuthenticationInfo

Authentication information for the operation.

AuthorizationInfo

Authorization information for the operation.

BigQueryAuditMetadata

Audit log format for BigQuery cloud audit logs metadata.

OrgPolicyViolationInfo

Represents OrgPolicy Violation information.

PolicyViolationInfo

Information related to policy violations for this request.

RequestMetadata

Metadata about the request.

ResourceLocation

Location information about a resource.

ServiceAccountDelegationInfo

Identity delegation history of an authenticated service account.

ViolationInfo

Provides information about the Policy violation info for this request.