Enum google_api_proto::google::cloud::audit::big_query_audit_metadata::Event

pub enum Event {

    JobInsertion(JobInsertion),
    JobChange(JobChange),
    JobDeletion(JobDeletion),
    DatasetCreation(DatasetCreation),
    DatasetChange(DatasetChange),
    DatasetDeletion(DatasetDeletion),
    TableCreation(TableCreation),
    TableChange(TableChange),
    TableDeletion(TableDeletion),
    TableDataRead(TableDataRead),
    TableDataChange(TableDataChange),
    ModelDeletion(ModelDeletion),
    ModelCreation(ModelCreation),
    ModelMetadataChange(ModelMetadataChange),
    ModelDataChange(ModelDataChange),
    ModelDataRead(ModelDataRead),
    RoutineCreation(RoutineCreation),
    RoutineChange(RoutineChange),
    RoutineDeletion(RoutineDeletion),
    RowAccessPolicyCreation(RowAccessPolicyCreation),
    RowAccessPolicyChange(RowAccessPolicyChange),
    RowAccessPolicyDeletion(RowAccessPolicyDeletion),
    UnlinkDataset(UnlinkDataset),

}

BigQuery event information.

Variants

JobInsertion(JobInsertion)

Job insertion event.

JobChange(JobChange)

Job state change event.

JobDeletion(JobDeletion)

Job deletion event.

DatasetCreation(DatasetCreation)

Dataset creation event.

DatasetChange(DatasetChange)

Dataset change event.

DatasetDeletion(DatasetDeletion)

Dataset deletion event.

TableCreation(TableCreation)

Table creation event.

TableChange(TableChange)

Table metadata change event.

TableDeletion(TableDeletion)

Table deletion event.

TableDataRead(TableDataRead)

Table data read event.

TableDataChange(TableDataChange)

Table data change event.

ModelDeletion(ModelDeletion)

Model deletion event.

ModelCreation(ModelCreation)

Model creation event.

ModelMetadataChange(ModelMetadataChange)

Model metadata change event.

ModelDataChange(ModelDataChange)

Model data change event.

ModelDataRead(ModelDataRead)

Model data read event.

RoutineCreation(RoutineCreation)

Routine creation event.

RoutineChange(RoutineChange)

Routine change event.

RoutineDeletion(RoutineDeletion)

Routine deletion event.

RowAccessPolicyCreation(RowAccessPolicyCreation)

Row access policy create event.

RowAccessPolicyChange(RowAccessPolicyChange)

Row access policy change event.

RowAccessPolicyDeletion(RowAccessPolicyDeletion)

Row access policy deletion event.

UnlinkDataset(UnlinkDataset)

Unlink linked dataset from its source dataset event

Implementations

impl Event

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Event>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.