Enum google_api_proto::google::cloud::audit::big_query_audit_metadata::QueryStatementType

#[repr(i32)]
pub enum QueryStatementType {

    Unspecified = 0,
    Select = 1,
    Assert = 23,
    Insert = 2,
    Update = 3,
    Delete = 4,
    Merge = 5,
    CreateTable = 6,
    CreateTableAsSelect = 7,
    CreateView = 8,
    CreateModel = 9,
    CreateMaterializedView = 13,
    CreateFunction = 14,
    CreateTableFunction = 56,
    CreateProcedure = 20,
    CreateRowAccessPolicy = 24,
    CreateSchema = 53,
    CreateSnapshotTable = 59,
    DropTable = 10,
    DropExternalTable = 33,
    DropView = 11,
    DropModel = 12,
    DropMaterializedView = 15,
    DropFunction = 16,
    DropProcedure = 21,
    DropSchema = 54,
    DropRowAccessPolicy = 25,
    DropSnapshotTable = 62,
    AlterTable = 17,
    AlterView = 18,
    AlterMaterializedView = 22,
    AlterSchema = 55,
    Script = 19,
    TruncateTable = 26,
    CreateExternalTable = 27,
    ExportData = 28,
    Call = 29,

}

Type of the statement (e.g. SELECT, INSERT, CREATE_TABLE, CREATE_MODEL..)

Variants

Unspecified = 0

Unknown.

Select = 1

SELECT … FROM <Table list> …

Assert = 23

ASSERT <condition> AS ‘description’

Insert = 2

INSERT INTO <Table> ….

Update = 3

UPDATE <Table> SET …

Delete = 4

DELETE <Table> …

Merge = 5

MERGE INTO <Table> ….

CreateTable = 6

CREATE TABLE <Table> <column list>

CreateTableAsSelect = 7

CREATE TABLE <Table> AS SELECT

CreateView = 8

CREATE VIEW <View>

CreateModel = 9

CREATE MODEL <Model> AS <Query>

CreateMaterializedView = 13

CREATE MATERIALIZED VIEW <View> AS …

CreateFunction = 14

CREATE FUNCTION <Function>(<Signature>) AS …

CreateTableFunction = 56

CREATE TABLE FUNCTION <Function>(<Signature>) AS …

CreateProcedure = 20

CREATE PROCEDURE <Procedure>

CreateRowAccessPolicy = 24

CREATE ROW ACCESS POLICY <RowAccessPolicy&gt ON <Table>

CreateSchema = 53

CREATE SCHEMA <Schema>

CreateSnapshotTable = 59

CREATE SNAPSHOT TABLE <Snapshot&gt CLONE <Table>

DropTable = 10

DROP TABLE <Table>

DropExternalTable = 33

DROP EXTERNAL TABLE <Table>

DropView = 11

DROP VIEW <View>

DropModel = 12

DROP MODEL <Model>

DropMaterializedView = 15

DROP MATERIALIZED VIEW <View>

DropFunction = 16

DROP FUNCTION <Function>

DropProcedure = 21

DROP PROCEDURE <Procedure>

DropSchema = 54

DROP SCHEMA <Schema>

DropRowAccessPolicy = 25

DROP ROW ACCESS POLICY <RowAccessPolicy&gt ON <Table> DROP ALL ROW ACCESS POLICIES ON ON <Table>

DropSnapshotTable = 62

DROP SNAPSHOT TABLE <Snapshot>

AlterTable = 17

ALTER TABLE <Table>

AlterView = 18

ALTER VIEW <View>

AlterMaterializedView = 22

ALTER MATERIALIZED_VIEW <view>

AlterSchema = 55

ALTER SCHEMA <Schema>

Script = 19

Script

TruncateTable = 26

TRUNCATE TABLE <Table>

CreateExternalTable = 27

CREATE EXTERNAL TABLE <TABLE>

ExportData = 28

EXPORT DATA;

Call = 29

CALL <stored procedure>

Implementations

impl QueryStatementType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of QueryStatementType.

pub fn from_i32(value: i32) -> Option<QueryStatementType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a QueryStatementType, or None if value is not a valid variant.

impl QueryStatementType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.