Module google_api_proto::google::cloud::baremetalsolution::v2

Modules

bare_metal_solution_client

Generated client implementations.

instance

Nested message and enum types in Instance.

instance_config

Nested message and enum types in InstanceConfig.

logical_interface

Nested message and enum types in LogicalInterface.

lun

Nested message and enum types in Lun.

network

Nested message and enum types in Network.

network_config

Nested message and enum types in NetworkConfig.

nfs_share

Nested message and enum types in NfsShare.

provisioning_config

Nested message and enum types in ProvisioningConfig.

provisioning_quota

Nested message and enum types in ProvisioningQuota.

server_network_template

Nested message and enum types in ServerNetworkTemplate.

volume

Nested message and enum types in Volume.

volume_config

Nested message and enum types in VolumeConfig.

volume_snapshot

Nested message and enum types in VolumeSnapshot.

vrf

Nested message and enum types in VRF.

Structs

CreateNfsShareRequest

Message for creating an NFS share.

CreateProvisioningConfigRequest

Request for CreateProvisioningConfig.

CreateSshKeyRequest

Message for registering a public SSH key in a project.

CreateVolumeSnapshotRequest

Message for creating a volume snapshot.

DeleteNfsShareRequest

Message for deleting an NFS share.

DeleteSshKeyRequest

Message for deleting an SSH key from a project.

DeleteVolumeSnapshotRequest

Message for deleting named Volume snapshot.

DetachLunRequest

Message for detach specific LUN from an Instance.

DisableInteractiveSerialConsoleRequest

Message for disabling the interactive serial console on an instance.

DisableInteractiveSerialConsoleResponse

Message for response of DisableInteractiveSerialConsole.

EnableInteractiveSerialConsoleRequest

Message for enabling the interactive serial console on an instance.

EnableInteractiveSerialConsoleResponse

Message for response of EnableInteractiveSerialConsole.

EvictLunRequest

Request for skip lun cooloff and delete it.

EvictVolumeRequest

Request for skip volume cooloff and delete it.

GetInstanceRequest

Message for requesting server information.

GetLunRequest

Message for requesting storage lun information.

GetNetworkRequest

Message for requesting network information.

GetNfsShareRequest

Message for requesting NFS share information.

GetProvisioningConfigRequest

Request for GetProvisioningConfig.

GetVolumeRequest

Message for requesting storage volume information.

GetVolumeSnapshotRequest

Message for requesting volume snapshot information.

Instance

A server.

InstanceConfig

Configuration parameters for a new instance.

InstanceQuota

A resource budget.

ListInstancesRequest

Message for requesting the list of servers.

ListInstancesResponse

Response message for the list of servers.

ListLunsRequest

Message for requesting a list of storage volume luns.

ListLunsResponse

Response message containing the list of storage volume luns.

ListNetworkUsageRequest

Request to get networks with IPs.

ListNetworkUsageResponse

Response with Networks with IPs

ListNetworksRequest

Message for requesting a list of networks.

ListNetworksResponse

Response message containing the list of networks.

ListNfsSharesRequest

Message for requesting a list of NFS shares.

ListNfsSharesResponse

Response message containing the list of NFS shares.

ListOsImagesRequest

Request for getting all available OS images.

ListOsImagesResponse

Request for getting all available OS images.

ListProvisioningQuotasRequest

Message for requesting the list of provisioning quotas.

ListProvisioningQuotasResponse

Response message for the list of provisioning quotas.

ListSshKeysRequest

Message for listing the public SSH keys in a project.

ListSshKeysResponse

Message for response of ListSSHKeys.

ListVolumeSnapshotsRequest

Message for requesting a list of volume snapshots.

ListVolumeSnapshotsResponse

Response message containing the list of volume snapshots.

ListVolumesRequest

Message for requesting a list of storage volumes.

ListVolumesResponse

Response message containing the list of storage volumes.

LogicalInterface

Each logical interface represents a logical abstraction of the underlying physical interface (for eg. bond, nic) of the instance. Each logical interface can effectively map to multiple network-IP pairs and still be mapped to one underlying physical interface.

Lun

A storage volume logical unit number (LUN).

Network

A Network.

NetworkAddressReservation

A reservation of one or more addresses in a network.

NetworkConfig

Configuration parameters for a new network.

NetworkMountPoint

Mount point for a network.

NetworkUsage

Network with all used IP addresses.

NfsShare

An NFS share.

OperationMetadata

Represents the metadata from a long-running operation.

OsImage

Operation System image.

ProvisioningConfig

A provisioning configuration.

ProvisioningQuota

A provisioning quota for a given project.

RenameInstanceRequest

Message requesting rename of a server.

RenameNetworkRequest

Message requesting rename of a server.

RenameNfsShareRequest

Message requesting rename of a server.

RenameVolumeRequest

Message requesting rename of a server.

ResetInstanceRequest

Message requesting to reset a server.

ResetInstanceResponse

Response message from resetting a server.

ResizeVolumeRequest

Request for emergency resize Volume.

RestoreVolumeSnapshotRequest

Message for restoring a volume snapshot.

ServerNetworkTemplate

Network template.

SshKey

An SSH key, used for authorizing with the interactive serial console feature.

StartInstanceRequest

Message requesting to start a server.

StartInstanceResponse

Response message from starting a server.

StopInstanceRequest

Message requesting to stop a server.

StopInstanceResponse

Response message from stopping a server.

SubmitProvisioningConfigRequest

Request for SubmitProvisioningConfig.

SubmitProvisioningConfigResponse

Response for SubmitProvisioningConfig.

UpdateInstanceRequest

Message requesting to updating a server.

UpdateNetworkRequest

Message requesting to updating a network.

UpdateNfsShareRequest

Message requesting to updating an NFS share.

UpdateProvisioningConfigRequest

Message for updating a ProvisioningConfig.

UpdateVolumeRequest

Message for updating a volume.

Volume

A storage volume.

VolumeConfig

Configuration parameters for a new volume.

VolumeSnapshot

A snapshot of a volume. Only boot volumes can have snapshots.

Vrf

A network VRF.

Enums

VolumePerformanceTier

Performance tier of the Volume.

WorkloadProfile

The possible values for a workload profile.