Module google_api_proto::google::cloud::batch::v1

Modules

allocation_policy

Nested message and enum types in AllocationPolicy.

batch_service_client

Generated client implementations.

environment

Nested message and enum types in Environment.

job_notification

Nested message and enum types in JobNotification.

job_status

Nested message and enum types in JobStatus.

lifecycle_policy

Nested message and enum types in LifecyclePolicy.

logs_policy

Nested message and enum types in LogsPolicy.

runnable

Nested message and enum types in Runnable.

task_group

Nested message and enum types in TaskGroup.

task_status

Nested message and enum types in TaskStatus.

volume

Nested message and enum types in Volume.

Structs

AllocationPolicy

A Job’s resource allocation policy describes when, where, and how compute resources should be allocated for the Job.

ComputeResource

Compute resource requirements.

CreateJobRequest

CreateJob Request.

DeleteJobRequest

DeleteJob Request.

Environment

An Environment describes a collection of environment variables to set when executing Tasks.

Gcs

Represents a Google Cloud Storage volume.

GetJobRequest

GetJob Request.

GetTaskRequest

Request for a single Task by name.

Job

The Cloud Batch Job description.

JobNotification

Notification configurations.

JobStatus

Job status.

LifecyclePolicy

LifecyclePolicy describes how to deal with task failures based on different conditions.

ListJobsRequest

ListJob Request.

ListJobsResponse

ListJob Response.

ListTasksRequest

ListTasks Request.

ListTasksResponse

ListTasks Response.

LogsPolicy

LogsPolicy describes how outputs from a Job’s Tasks (stdout/stderr) will be preserved.

Nfs

Represents an NFS volume.

OperationMetadata

Represents the metadata of the long-running operation.

Runnable

Runnable describes instructions for executing a specific script or container as part of a Task.

ServiceAccount

Carries information about a Google Cloud service account.

StatusEvent

Status event

Task

A Cloud Batch task.

TaskExecution

This Task Execution field includes detail information for task execution procedures, based on StatusEvent types.

TaskGroup

A TaskGroup defines one or more Tasks that all share the same TaskSpec.

TaskSpec

Spec of a task

TaskStatus

Status of a task

Volume

Volume describes a volume and parameters for it to be mounted to a VM.