Module google_api_proto::google::cloud::batch::v1alpha

Modules

allocation_policy

Nested message and enum types in AllocationPolicy.

batch_service_client

Generated client implementations.

environment

Nested message and enum types in Environment.

job

Nested message and enum types in Job.

job_dependency

Nested message and enum types in JobDependency.

job_notification

Nested message and enum types in JobNotification.

job_status

Nested message and enum types in JobStatus.

lifecycle_policy

Nested message and enum types in LifecyclePolicy.

logs_policy

Nested message and enum types in LogsPolicy.

resource_allowance

Nested message and enum types in ResourceAllowance.

runnable

Nested message and enum types in Runnable.

task_group

Nested message and enum types in TaskGroup.

task_status

Nested message and enum types in TaskStatus.

usage_resource_allowance_spec

Nested message and enum types in UsageResourceAllowanceSpec.

usage_resource_allowance_status

Nested message and enum types in UsageResourceAllowanceStatus.

volume

Nested message and enum types in Volume.

Structs

AllocationPolicy

A Job’s resource allocation policy describes when, where, and how compute resources should be allocated for the Job.

CancelJobRequest

CancelJob Request.

CancelJobResponse

Response to the CancelJob request.

ComputeResource

Compute resource requirements.

CreateJobRequest

CreateJob Request.

CreateResourceAllowanceRequest

CreateResourceAllowance Request.

DeleteJobRequest

DeleteJob Request.

DeleteResourceAllowanceRequest

DeleteResourceAllowance Request.

Environment

An Environment describes a collection of environment variables to set when executing Tasks.

Gcs

Represents a Google Cloud Storage volume.

GetJobRequest

GetJob Request.

GetResourceAllowanceRequest

GetResourceAllowance Request.

GetTaskRequest

Request for a single Task by name.

Job

The Cloud Batch Job description.

JobDependency

JobDependency describes the state of other Jobs that the start of this Job depends on. All dependent Jobs must have been submitted in the same region.

JobNotification

Notification configurations.

JobStatus

Job status.

LifecyclePolicy

LifecyclePolicy describes how to deal with task failures based on different conditions.

ListJobsRequest

ListJob Request.

ListJobsResponse

ListJob Response.

ListResourceAllowancesRequest

ListResourceAllowances Request.

ListResourceAllowancesResponse

ListResourceAllowances Response.

ListTasksRequest

ListTasks Request.

ListTasksResponse

ListTasks Response.

LogsPolicy

LogsPolicy describes how outputs from a Job’s Tasks (stdout/stderr) will be preserved.

Nfs

Represents an NFS volume.

Notification

Notification on resource state change.

OperationMetadata

Represents the metadata of the long-running operation.

Pd

Deprecated: please use device_name instead.

ResourceAllowance

The Resource Allowance description for Cloud Batch. Only one Resource Allowance is supported now under a specific location and project.

ResourceUsage

ResourceUsage describes the resource usage of the job.

Runnable

Runnable describes instructions for executing a specific script or container as part of a Task.

ServiceAccount

Carries information about a Google Cloud service account.

StatusEvent

Status event.

Task

A Cloud Batch task.

TaskExecution

This Task Execution field includes detail information for task execution procedures, based on StatusEvent types.

TaskGroup

A TaskGroup defines one or more Tasks that all share the same TaskSpec.

TaskResourceUsage

TaskResourceUsage describes the resource usage of the task.

TaskSpec

Spec of a task

TaskStatus

Status of a task.

UpdateJobRequest

UpdateJob Request.

UpdateResourceAllowanceRequest

UpdateResourceAllowance Request.

UsageResourceAllowance

UsageResourceAllowance describes the detail of usage resource allowance.

UsageResourceAllowanceSpec

Spec of a usage ResourceAllowance.

UsageResourceAllowanceStatus

Status of a usage ResourceAllowance.

Volume

Volume describes a volume and parameters for it to be mounted to a VM.

Enums

CalendarPeriod

A CalendarPeriod represents the abstract concept of a time period that has a canonical start. All calendar times begin at 12 AM US and Canadian Pacific Time (UTC-8).

ResourceAllowanceState

ResourceAllowance valid state.