Module google_api_proto::google::cloud::batch::v1alpha::allocation_policy

Nested message and enum types in AllocationPolicy.

Modules

attached_disk

Nested message and enum types in AttachedDisk.

disk

Nested message and enum types in Disk.

instance_policy_or_template

Nested message and enum types in InstancePolicyOrTemplate.

Structs

Accelerator

Accelerator describes Compute Engine accelerators to be attached to the VM.

AttachedDisk

A new or an existing persistent disk (PD) or a local ssd attached to a VM instance.

Disk

A new persistent disk or a local ssd. A VM can only have one local SSD setting but multiple local SSD partitions. See https://cloud.google.com/compute/docs/disks#pdspecs and https://cloud.google.com/compute/docs/disks#localssds.

InstancePolicy

InstancePolicy describes an instance type and resources attached to each VM created by this InstancePolicy.

InstancePolicyOrTemplate

InstancePolicyOrTemplate lets you define the type of resources to use for this job either with an InstancePolicy or an instance template. If undefined, Batch picks the type of VM to use and doesn’t include optional VM resources such as GPUs and extra disks.

LocationPolicy

NetworkInterface

A network interface.

NetworkPolicy

NetworkPolicy describes VM instance network configurations.

PlacementPolicy

PlacementPolicy describes a group placement policy for the VMs controlled by this AllocationPolicy.

Enums

ProvisioningModel

Compute Engine VM instance provisioning model.