Module google_api_proto::google::cloud::batch::v1alpha1

Modules

allocation_policy

Nested message and enum types in AllocationPolicy.

batch_service_client

Generated client implementations.

job

Nested message and enum types in Job.

job_dependency

Nested message and enum types in JobDependency.

job_status

Nested message and enum types in JobStatus.

lifecycle_policy

Nested message and enum types in LifecyclePolicy.

logs_policy

Nested message and enum types in LogsPolicy.

runnable

Nested message and enum types in Runnable.

task_group

Nested message and enum types in TaskGroup.

task_status

Nested message and enum types in TaskStatus.

volume

Nested message and enum types in Volume.

Structs

AllocationPolicy

A Job’s resource allocation policy describes when, where, and how compute resources should be allocated for the Job.

ComputeResource

Compute resource requirements

CreateJobRequest

CreateJob Request.

DeleteJobRequest

DeleteJob Request.

Gcs

Represents a Google Cloud Storage volume source config.

GetJobRequest

GetJob Request.

GetTaskRequest

Request for a single Task by name.

Job

The Cloud Batch Job description.

JobDependency

JobDependency describes the state of other Jobs that the start of this Job depends on. All dependent Jobs must have been submitted in the same region.

JobNotification

Job notification.

JobStatus

Job status.

LifecyclePolicy

LifecyclePolicy describes how to deal with task failures based on different conditions.

ListJobsRequest

ListJob Request.

ListJobsResponse

ListJob Response.

ListTasksRequest

ListTasks Request.

ListTasksResponse

ListAssignedTasks Response.

LogsPolicy

LogsPolicy describes how outputs from a Job’s Tasks (stdout/stderr) will be preserved.

Nfs

Represents an NFS server and remote path: :<remote_path>

OperationMetadata

Represents the metadata of the long-running operation.

Pd

Represents a GCP persistent disk

Runnable

Runnable describes instructions for executing a specific script or container as part of a Task.

StatusEvent

Status event

Task

A Cloud Batch task.

TaskExecution

This Task Execution field includes detail information for task execution procedures, based on StatusEvent types.

TaskGroup

A TaskGroup contains one or multiple Tasks that share the same Runnable but with different runtime parameters.

TaskSpec

Spec of a task

TaskStatus

Status of a task

Volume

Volume and mount parameters to be associated with a TaskSpec. A TaskSpec might describe zero, one, or multiple volumes to be mounted as part of the task.