Module google_api_proto::google::cloud::batch::v1alpha1::allocation_policy

Nested message and enum types in AllocationPolicy.

Structs

InstancePolicy

InstancePolicy describes what instances should be created for the Job.

LocationPolicy

Be consistent with LocationPolicy in //cloud/cluster/api/mixter_instances.proto.

NetworkInterface

A network interface.

NetworkPolicy

NetworkPolicy describes network configurations for instances created for the job.

Enums

ProvisioningModel

Compute Engine VM instance provisioning model.