Module google_api_proto::google::cloud::beyondcorp::appconnectors::v1

Modules

app_connector

Nested message and enum types in AppConnector.

app_connectors_service_client

Generated client implementations.

notification_config

Nested message and enum types in NotificationConfig.

Structs

AppConnector

A BeyondCorp connector resource that represents an application facing component deployed proximal to and with direct access to the application instances. It is used to establish connectivity between the remote enterprise environment and GCP. It initiates connections to the applications and can proxy the data from users over the connection.

AppConnectorInstanceConfig

AppConnectorInstanceConfig defines the instance config of a AppConnector.

AppConnectorOperationMetadata

Represents the metadata of the long-running operation.

CreateAppConnectorRequest

Request message for BeyondCorp.CreateAppConnector.

DeleteAppConnectorRequest

Request message for BeyondCorp.DeleteAppConnector.

GetAppConnectorRequest

Request message for BeyondCorp.GetAppConnector.

ImageConfig

ImageConfig defines the control plane images to run.

ListAppConnectorsRequest

Request message for BeyondCorp.ListAppConnectors.

ListAppConnectorsResponse

Response message for BeyondCorp.ListAppConnectors.

NotificationConfig

NotificationConfig defines the mechanisms to notify instance agent.

ReportStatusRequest

Request report the connector status.

ResourceInfo

ResourceInfo represents the information/status of an app connector resource. Such as:

UpdateAppConnectorRequest

Request message for BeyondCorp.UpdateAppConnector.

Enums

HealthStatus

HealthStatus represents the health status.