Module google_api_proto::google::cloud::bigquery::analyticshub::v1

Modules

analytics_hub_service_client

Generated client implementations.

listing

Nested message and enum types in Listing.

sharing_environment_config

Nested message and enum types in SharingEnvironmentConfig.

subscribe_listing_request

Nested message and enum types in SubscribeListingRequest.

subscription

Nested message and enum types in Subscription.

Structs

CreateDataExchangeRequest

Message for creating a data exchange.

CreateListingRequest

Message for creating a listing.

DataExchange

A data exchange is a container that lets you share data. Along with the descriptive information about the data exchange, it contains listings that reference shared datasets.

DataProvider

Contains details of the data provider.

DeleteDataExchangeRequest

Message for deleting a data exchange.

DeleteListingRequest

Message for deleting a listing.

DeleteSubscriptionRequest

Message for deleting a subscription.

DestinationDataset

Defines the destination bigquery dataset.

DestinationDatasetReference

Contains the reference that identifies a destination bigquery dataset.

GetDataExchangeRequest

Message for getting a data exchange.

GetListingRequest

Message for getting a listing.

GetSubscriptionRequest

Message for getting a subscription.

ListDataExchangesRequest

Message for requesting the list of data exchanges.

ListDataExchangesResponse

Message for response to the list of data exchanges.

ListListingsRequest

Message for requesting the list of listings.

ListListingsResponse

Message for response to the list of Listings.

ListOrgDataExchangesRequest

Message for requesting the list of data exchanges from projects in an organization and location.

ListOrgDataExchangesResponse

Message for response to listing data exchanges in an organization and location.

ListSharedResourceSubscriptionsRequest

Message for listing subscriptions of a shared resource.

ListSharedResourceSubscriptionsResponse

Message for response to the listing of shared resource subscriptions.

ListSubscriptionsRequest

Message for listing subscriptions.

ListSubscriptionsResponse

Message for response to the listing of subscriptions.

Listing

A listing is what gets published into a data exchange that a subscriber can subscribe to. It contains a reference to the data source along with descriptive information that will help subscribers find and subscribe the data.

OperationMetadata

Represents the metadata of a long-running operation in Analytics Hub.

Publisher

Contains details of the listing publisher.

RefreshSubscriptionRequest

Message for refreshing a subscription.

RefreshSubscriptionResponse

Message for response when you refresh a subscription.

RevokeSubscriptionRequest

Message for revoking a subscription.

RevokeSubscriptionResponse

Message for response when you revoke a subscription.

SharingEnvironmentConfig

Sharing environment is a behavior model for sharing data within a data exchange. This option is configurable for a data exchange.

SubscribeDataExchangeRequest

Message for subscribing to a Data Exchange.

SubscribeDataExchangeResponse

Message for response when you subscribe to a Data Exchange.

SubscribeListingRequest

Message for subscribing to a listing.

SubscribeListingResponse

Message for response when you subscribe to a listing.

Subscription

A subscription represents a subscribers’ access to a particular set of published data. It contains references to associated listings, data exchanges, and linked datasets.

UpdateDataExchangeRequest

Message for updating a data exchange.

UpdateListingRequest

Message for updating a Listing.

Enums

DiscoveryType

Specifies the type of discovery on the discovery page. Note that this does not control the visibility of the exchange/listing which is defined by IAM permission.