Struct google_api_proto::google::cloud::bigquery::analyticshub::v1::analytics_hub_service_client::AnalyticsHubServiceClient

pub struct AnalyticsHubServiceClient<T> { /* private fields */ }

The AnalyticsHubService API facilitates data sharing within and across organizations. It allows data providers to publish listings that reference shared datasets. With Analytics Hub, users can discover and search for listings that they have access to. Subscribers can view and subscribe to listings. When you subscribe to a listing, Analytics Hub creates a linked dataset in your project.

Implementations

impl<T> AnalyticsHubServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> AnalyticsHubServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_data_exchanges( &mut self, request: impl IntoRequest<ListDataExchangesRequest>, ) -> Result<Response<ListDataExchangesResponse>, Status>

Lists all data exchanges in a given project and location.

pub async fn list_org_data_exchanges( &mut self, request: impl IntoRequest<ListOrgDataExchangesRequest>, ) -> Result<Response<ListOrgDataExchangesResponse>, Status>

Lists all data exchanges from projects in a given organization and location.

pub async fn get_data_exchange( &mut self, request: impl IntoRequest<GetDataExchangeRequest>, ) -> Result<Response<DataExchange>, Status>

Gets the details of a data exchange.

pub async fn create_data_exchange( &mut self, request: impl IntoRequest<CreateDataExchangeRequest>, ) -> Result<Response<DataExchange>, Status>

Creates a new data exchange.

pub async fn update_data_exchange( &mut self, request: impl IntoRequest<UpdateDataExchangeRequest>, ) -> Result<Response<DataExchange>, Status>

Updates an existing data exchange.

pub async fn delete_data_exchange( &mut self, request: impl IntoRequest<DeleteDataExchangeRequest>, ) -> Result<Response<()>, Status>

Deletes an existing data exchange.

pub async fn list_listings( &mut self, request: impl IntoRequest<ListListingsRequest>, ) -> Result<Response<ListListingsResponse>, Status>

Lists all listings in a given project and location.

pub async fn get_listing( &mut self, request: impl IntoRequest<GetListingRequest>, ) -> Result<Response<Listing>, Status>

Gets the details of a listing.

pub async fn create_listing( &mut self, request: impl IntoRequest<CreateListingRequest>, ) -> Result<Response<Listing>, Status>

Creates a new listing.

pub async fn update_listing( &mut self, request: impl IntoRequest<UpdateListingRequest>, ) -> Result<Response<Listing>, Status>

Updates an existing listing.

pub async fn delete_listing( &mut self, request: impl IntoRequest<DeleteListingRequest>, ) -> Result<Response<()>, Status>

Deletes a listing.

pub async fn subscribe_listing( &mut self, request: impl IntoRequest<SubscribeListingRequest>, ) -> Result<Response<SubscribeListingResponse>, Status>

Subscribes to a listing.

Currently, with Analytics Hub, you can create listings that reference only BigQuery datasets. Upon subscription to a listing for a BigQuery dataset, Analytics Hub creates a linked dataset in the subscriber’s project.

pub async fn subscribe_data_exchange( &mut self, request: impl IntoRequest<SubscribeDataExchangeRequest>, ) -> Result<Response<Operation>, Status>

Creates a Subscription to a Data Exchange. This is a long-running operation as it will create one or more linked datasets.

pub async fn refresh_subscription( &mut self, request: impl IntoRequest<RefreshSubscriptionRequest>, ) -> Result<Response<Operation>, Status>

Refreshes a Subscription to a Data Exchange. A Data Exchange can become stale when a publisher adds or removes data. This is a long-running operation as it may create many linked datasets.

pub async fn get_subscription( &mut self, request: impl IntoRequest<GetSubscriptionRequest>, ) -> Result<Response<Subscription>, Status>

Gets the details of a Subscription.

pub async fn list_subscriptions( &mut self, request: impl IntoRequest<ListSubscriptionsRequest>, ) -> Result<Response<ListSubscriptionsResponse>, Status>

Lists all subscriptions in a given project and location.

pub async fn list_shared_resource_subscriptions( &mut self, request: impl IntoRequest<ListSharedResourceSubscriptionsRequest>, ) -> Result<Response<ListSharedResourceSubscriptionsResponse>, Status>

Lists all subscriptions on a given Data Exchange or Listing.

pub async fn revoke_subscription( &mut self, request: impl IntoRequest<RevokeSubscriptionRequest>, ) -> Result<Response<RevokeSubscriptionResponse>, Status>

Revokes a given subscription.

pub async fn delete_subscription( &mut self, request: impl IntoRequest<DeleteSubscriptionRequest>, ) -> Result<Response<Operation>, Status>

Deletes a subscription.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets the IAM policy.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Sets the IAM policy.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest>, ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the permissions that a caller has.