Struct google_api_proto::google::cloud::bigquery::biglake::v1alpha1::metastore_service_client::MetastoreServiceClient

pub struct MetastoreServiceClient<T> { /* private fields */ }

BigLake Metastore is a serverless, highly available, multi-tenant runtime metastore for Google Cloud Data Analytics products.

The BigLake Metastore API defines the following resource model:

A collection of Google Cloud projects: /projects/*

Each project has a collection of available locations: /locations/*

Each location has a collection of catalogs: /catalogs/*

Each catalog has a collection of databases: /databases/*

Each database has a collection of tables: /tables/*

Implementations

impl<T> MetastoreServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> MetastoreServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_catalog( &mut self, request: impl IntoRequest<CreateCatalogRequest> ) -> Result<Response<Catalog>, Status>

Creates a new catalog.

pub async fn delete_catalog( &mut self, request: impl IntoRequest<DeleteCatalogRequest> ) -> Result<Response<Catalog>, Status>

Deletes an existing catalog specified by the catalog ID.

pub async fn get_catalog( &mut self, request: impl IntoRequest<GetCatalogRequest> ) -> Result<Response<Catalog>, Status>

Gets the catalog specified by the resource name.

pub async fn list_catalogs( &mut self, request: impl IntoRequest<ListCatalogsRequest> ) -> Result<Response<ListCatalogsResponse>, Status>

List all catalogs in a specified project.

pub async fn create_database( &mut self, request: impl IntoRequest<CreateDatabaseRequest> ) -> Result<Response<Database>, Status>

Creates a new database.

pub async fn delete_database( &mut self, request: impl IntoRequest<DeleteDatabaseRequest> ) -> Result<Response<Database>, Status>

Deletes an existing database specified by the database ID.

pub async fn update_database( &mut self, request: impl IntoRequest<UpdateDatabaseRequest> ) -> Result<Response<Database>, Status>

Updates an existing database specified by the database ID.

pub async fn get_database( &mut self, request: impl IntoRequest<GetDatabaseRequest> ) -> Result<Response<Database>, Status>

Gets the database specified by the resource name.

pub async fn list_databases( &mut self, request: impl IntoRequest<ListDatabasesRequest> ) -> Result<Response<ListDatabasesResponse>, Status>

List all databases in a specified catalog.

pub async fn create_table( &mut self, request: impl IntoRequest<CreateTableRequest> ) -> Result<Response<Table>, Status>

Creates a new table.

pub async fn delete_table( &mut self, request: impl IntoRequest<DeleteTableRequest> ) -> Result<Response<Table>, Status>

Deletes an existing table specified by the table ID.

pub async fn update_table( &mut self, request: impl IntoRequest<UpdateTableRequest> ) -> Result<Response<Table>, Status>

Updates an existing table specified by the table ID.

pub async fn rename_table( &mut self, request: impl IntoRequest<RenameTableRequest> ) -> Result<Response<Table>, Status>

Renames an existing table specified by the table ID.

pub async fn get_table( &mut self, request: impl IntoRequest<GetTableRequest> ) -> Result<Response<Table>, Status>

Gets the table specified by the resource name.

pub async fn list_tables( &mut self, request: impl IntoRequest<ListTablesRequest> ) -> Result<Response<ListTablesResponse>, Status>

List all tables in a specified database.

pub async fn create_lock( &mut self, request: impl IntoRequest<CreateLockRequest> ) -> Result<Response<Lock>, Status>

Creates a new lock.

pub async fn delete_lock( &mut self, request: impl IntoRequest<DeleteLockRequest> ) -> Result<Response<()>, Status>

Deletes an existing lock specified by the lock ID.

pub async fn check_lock( &mut self, request: impl IntoRequest<CheckLockRequest> ) -> Result<Response<Lock>, Status>

Checks the state of a lock specified by the lock ID.

pub async fn list_locks( &mut self, request: impl IntoRequest<ListLocksRequest> ) -> Result<Response<ListLocksResponse>, Status>

List all locks in a specified database.