Module google_api_proto::google::cloud::bigquery::connection::v1

Modules

aws_properties

Nested message and enum types in AwsProperties.

cloud_sql_properties

Nested message and enum types in CloudSqlProperties.

connection

Nested message and enum types in Connection.

connection_service_client

Generated client implementations.

Structs

AwsAccessRole

Authentication method for Amazon Web Services (AWS) that uses Google owned Google service account to assume into customer’s AWS IAM Role.

AwsCrossAccountRole

Authentication method for Amazon Web Services (AWS) that uses Google owned AWS IAM user’s access key to assume into customer’s AWS IAM Role.

AwsProperties

Connection properties specific to Amazon Web Services (AWS).

AzureProperties

Container for connection properties specific to Azure.

CloudResourceProperties

Container for connection properties for delegation of access to GCP resources.

CloudSpannerProperties

Connection properties specific to Cloud Spanner.

CloudSqlCredential

Credential info for the Cloud SQL.

CloudSqlProperties

Connection properties specific to the Cloud SQL.

Connection

Configuration parameters to establish connection with an external data source, except the credential attributes.

CreateConnectionRequest

The request for [ConnectionService.CreateConnection][google.cloud.bigquery.connection.v1.ConnectionService.CreateConnection].

DeleteConnectionRequest

The request for [ConnectionService.DeleteConnectionRequest][].

GetConnectionRequest

The request for [ConnectionService.GetConnection][google.cloud.bigquery.connection.v1.ConnectionService.GetConnection].

ListConnectionsRequest

The request for [ConnectionService.ListConnections][google.cloud.bigquery.connection.v1.ConnectionService.ListConnections].

ListConnectionsResponse

The response for [ConnectionService.ListConnections][google.cloud.bigquery.connection.v1.ConnectionService.ListConnections].

MetastoreServiceConfig

Configuration of the Dataproc Metastore Service.

SalesforceDataCloudProperties

Connection properties specific to Salesforce DataCloud. This is intended for use only by Salesforce partner projects.

SparkHistoryServerConfig

Configuration of the Spark History Server.

SparkProperties

Container for connection properties to execute stored procedures for Apache Spark.

UpdateConnectionRequest

The request for [ConnectionService.UpdateConnection][google.cloud.bigquery.connection.v1.ConnectionService.UpdateConnection].