Module google_api_proto::google::cloud::bigquery::datatransfer::v1

Modules

data_source

Nested message and enum types in DataSource.

data_source_parameter

Nested message and enum types in DataSourceParameter.

data_transfer_service_client

Generated client implementations.

list_transfer_runs_request

Nested message and enum types in ListTransferRunsRequest.

start_manual_transfer_runs_request

Nested message and enum types in StartManualTransferRunsRequest.

transfer_config

Nested message and enum types in TransferConfig.

transfer_message

Nested message and enum types in TransferMessage.

transfer_run

Nested message and enum types in TransferRun.

Structs

CheckValidCredsRequest

A request to determine whether the user has valid credentials. This method is used to limit the number of OAuth popups in the user interface. The user id is inferred from the API call context. If the data source has the Google+ authorization type, this method returns false, as it cannot be determined whether the credentials are already valid merely based on the user id.

CheckValidCredsResponse

A response indicating whether the credentials exist and are valid.

CreateTransferConfigRequest

A request to create a data transfer configuration. If new credentials are needed for this transfer configuration, authorization info must be provided. If authorization info is provided, the transfer configuration will be associated with the user id corresponding to the authorization info. Otherwise, the transfer configuration will be associated with the calling user.

DataSource

Defines the properties and custom parameters for a data source.

DataSourceParameter

A parameter used to define custom fields in a data source definition.

DeleteTransferConfigRequest

A request to delete data transfer information. All associated transfer runs and log messages will be deleted as well.

DeleteTransferRunRequest

A request to delete data transfer run information.

EmailPreferences

Represents preferences for sending email notifications for transfer run events.

EncryptionConfiguration

Represents the encryption configuration for a transfer.

EnrollDataSourcesRequest

A request to enroll a set of data sources so they are visible in the BigQuery UI’s Transfer tab.

GetDataSourceRequest

A request to get data source info.

GetTransferConfigRequest

A request to get data transfer information.

GetTransferRunRequest

A request to get data transfer run information.

ListDataSourcesRequest

Request to list supported data sources and their data transfer settings.

ListDataSourcesResponse

Returns list of supported data sources and their metadata.

ListTransferConfigsRequest

A request to list data transfers configured for a BigQuery project.

ListTransferConfigsResponse

The returned list of pipelines in the project.

ListTransferLogsRequest

A request to get user facing log messages associated with data transfer run.

ListTransferLogsResponse

The returned list transfer run messages.

ListTransferRunsRequest

A request to list data transfer runs.

ListTransferRunsResponse

The returned list of pipelines in the project.

ScheduleOptions

Options customizing the data transfer schedule.

ScheduleTransferRunsRequest

A request to schedule transfer runs for a time range.

ScheduleTransferRunsResponse

A response to schedule transfer runs for a time range.

StartManualTransferRunsRequest

A request to start manual transfer runs.

StartManualTransferRunsResponse

A response to start manual transfer runs.

TransferConfig

Represents a data transfer configuration. A transfer configuration contains all metadata needed to perform a data transfer. For example, destination_dataset_id specifies where data should be stored. When a new transfer configuration is created, the specified destination_dataset_id is created when needed and shared with the appropriate data source service account.

TransferMessage

Represents a user facing message for a particular data transfer run.

TransferRun

Represents a data transfer run.

UnenrollDataSourcesRequest

A request to unenroll a set of data sources so they are no longer visible in the BigQuery UI’s Transfer tab.

UpdateTransferConfigRequest

A request to update a transfer configuration. To update the user id of the transfer configuration, authorization info needs to be provided.

UserInfo

Information about a user.

Enums

TransferState

Represents data transfer run state.

TransferType

DEPRECATED. Represents data transfer type.