Module google_api_proto::google::cloud::bigquery::migration::v2

Modules

dialect

Nested message and enum types in Dialect.

migration_service_client

Generated client implementations.

migration_subtask

Nested message and enum types in MigrationSubtask.

migration_task

Nested message and enum types in MigrationTask.

migration_workflow

Nested message and enum types in MigrationWorkflow.

name_mapping_key

Nested message and enum types in NameMappingKey.

teradata_dialect

Nested message and enum types in TeradataDialect.

translation_config_details

Nested message and enum types in TranslationConfigDetails.

typed_value

Nested message and enum types in TypedValue.

Structs

AzureSynapseDialect

The dialect definition for Azure Synapse.

BigQueryDialect

The dialect definition for BigQuery.

CreateMigrationWorkflowRequest

Request to create a migration workflow resource.

DeleteMigrationWorkflowRequest

A request to delete a previously created migration workflow.

Dialect

The possible dialect options for translation.

ErrorDetail

Provides details for errors, e.g. issues that where encountered when processing a subtask.

ErrorLocation

Holds information about where the error is located.

GetMigrationSubtaskRequest

A request to get a previously created migration subtasks.

GetMigrationWorkflowRequest

A request to get a previously created migration workflow.

HiveQlDialect

The dialect definition for HiveQL.

ListMigrationSubtasksRequest

A request to list previously created migration subtasks.

ListMigrationSubtasksResponse

Response object for a ListMigrationSubtasks call.

ListMigrationWorkflowsRequest

A request to list previously created migration workflows.

ListMigrationWorkflowsResponse

Response object for a ListMigrationWorkflows call.

MigrationSubtask

A subtask for a migration which carries details about the configuration of the subtask. The content of the details should not matter to the end user, but is a contract between the subtask creator and subtask worker.

MigrationTask

A single task for a migration which has details about the configuration of the task.

MigrationWorkflow

A migration workflow which specifies what needs to be done for an EDW migration.

MySqlDialect

The dialect definition for MySQL.

NameMappingKey

The potential components of a full name mapping that will be mapped during translation in the source data warehouse.

NameMappingValue

The potential components of a full name mapping that will be mapped during translation in the target data warehouse.

NetezzaDialect

The dialect definition for Netezza.

ObjectNameMapping

Represents a key-value pair of NameMappingKey to NameMappingValue to represent the mapping of SQL names from the input value to desired output.

ObjectNameMappingList

Represents a map of name mappings using a list of key:value proto messages of existing name to desired output name.

OracleDialect

The dialect definition for Oracle.

Point

A single data point in a time series.

PostgresqlDialect

The dialect definition for Postgresql.

PrestoDialect

The dialect definition for Presto.

RedshiftDialect

The dialect definition for Redshift.

ResourceErrorDetail

Provides details for errors and the corresponding resources.

SnowflakeDialect

The dialect definition for Snowflake.

SourceEnv

Represents the default source environment values for the translation.

SparkSqlDialect

The dialect definition for SparkSQL.

SqlServerDialect

The dialect definition for SQL Server.

StartMigrationWorkflowRequest

A request to start a previously created migration workflow.

TeradataDialect

The dialect definition for Teradata.

TimeInterval

A time interval extending just after a start time through an end time. If the start time is the same as the end time, then the interval represents a single point in time.

TimeSeries

The metrics object for a SubTask.

TranslationConfigDetails

The translation config to capture necessary settings for a translation task and subtask.

TypedValue

A single strongly-typed value.

VerticaDialect

The dialect definition for Vertica.