Module google_api_proto::google::cloud::bigquery::migration::v2alpha

Modules

identifier_settings

Nested message and enum types in IdentifierSettings.

migration_service_client

Generated client implementations.

migration_subtask

Nested message and enum types in MigrationSubtask.

migration_task

Nested message and enum types in MigrationTask.

migration_task_orchestration_result

Nested message and enum types in MigrationTaskOrchestrationResult.

migration_workflow

Nested message and enum types in MigrationWorkflow.

sql_translation_error

Nested message and enum types in SqlTranslationError.

sql_translation_service_client

Generated client implementations.

translate_query_request

Nested message and enum types in TranslateQueryRequest.

translation_task_details

Nested message and enum types in TranslationTaskDetails.

typed_value

Nested message and enum types in TypedValue.

Structs

AssessmentOrchestrationResultDetails

Details for an assessment task orchestration result.

AssessmentTaskDetails

Assessment task config.

BteqOptions

BTEQ translation task related settings.

CreateMigrationWorkflowRequest

Request to create a migration workflow resource.

DatasetReference

Reference to a BigQuery dataset.

DeleteMigrationWorkflowRequest

A request to delete a previously created migration workflow.

ErrorDetail

Provides details for errors, e.g. issues that where encountered when processing a subtask.

ErrorLocation

Holds information about where the error is located.

Filter

The filter applied to fields of translation details.

GetMigrationSubtaskRequest

A request to get a previously created migration subtasks.

GetMigrationWorkflowRequest

A request to get a previously created migration workflow.

IdentifierSettings

Settings related to SQL identifiers.

ListMigrationSubtasksRequest

A request to list previously created migration subtasks.

ListMigrationSubtasksResponse

Response object for a ListMigrationSubtasks call.

ListMigrationWorkflowsRequest

A request to list previously created migration workflows.

ListMigrationWorkflowsResponse

Response object for a ListMigrationWorkflows call.

MigrationSubtask

A subtask for a migration which carries details about the configuration of the subtask. The content of the details should not matter to the end user, but is a contract between the subtask creator and subtask worker.

MigrationTask

A single task for a migration which has details about the configuration of the task.

MigrationTaskOrchestrationResult

Additional information from the orchestrator when it is done with the task orchestration.

MigrationWorkflow

A migration workflow which specifies what needs to be done for an EDW migration.

Point

A single data point in a time series.

ResourceErrorDetail

Provides details for errors and the corresponding resources.

SqlTranslationError

The detailed error object if the SQL translation job fails.

SqlTranslationErrorDetail

Structured error object capturing the error message and the location in the source text where the error occurs.

SqlTranslationWarning

The detailed warning object if the SQL translation job is completed but not semantically correct.

StartMigrationWorkflowRequest

A request to start a previously created migration workflow.

TeradataOptions

Teradata SQL specific translation task related settings.

TimeInterval

A time interval extending just after a start time through an end time. If the start time is the same as the end time, then the interval represents a single point in time.

TimeSeries

The metrics object for a SubTask.

TranslateQueryRequest

The request of translating a SQL query to Standard SQL.

TranslateQueryResponse

The response of translating a SQL query to Standard SQL.

TranslationFileMapping

Mapping between an input and output file to be translated in a subtask.

TranslationTaskDetails

The translation task config to capture necessary settings for a translation task and subtask.

TypedValue

A single strongly-typed value.