Struct google_api_proto::google::cloud::bigquery::migration::v2alpha::TranslationTaskDetails

pub struct TranslationTaskDetails {
    pub input_path: String,
    pub output_path: String,
    pub file_paths: Vec<TranslationFileMapping>,
    pub schema_path: String,
    pub file_encoding: i32,
    pub identifier_settings: Option<IdentifierSettings>,
    pub special_token_map: BTreeMap<String, i32>,
    pub filter: Option<Filter>,
    pub translation_exception_table: String,
    pub language_options: Option<LanguageOptions>,
}

The translation task config to capture necessary settings for a translation task and subtask.

Fields
input_path: String

The Cloud Storage path for translation input files.

output_path: String

The Cloud Storage path for translation output files.

file_paths: Vec<TranslationFileMapping>

Cloud Storage files to be processed for translation.

schema_path: String

The Cloud Storage path to DDL files as table schema to assist semantic translation.

file_encoding: i32

The file encoding type.

identifier_settings: Option<IdentifierSettings>

The settings for SQL identifiers.

special_token_map: BTreeMap<String, i32>

The map capturing special tokens to be replaced during translation. The key is special token in string. The value is the token data type. This is used to translate SQL query template which contains special token as place holder. The special token makes a query invalid to parse. This map will be applied to annotate those special token with types to let parser understand how to parse them into proper structure with type information.

filter: Option<Filter>

The filter applied to translation details.

translation_exception_table: String

Specifies the exact name of the bigquery table (“dataset.table”) to be used for surfacing raw translation errors. If the table does not exist, we will create it. If it already exists and the schema is the same, we will re-use. If the table exists and the schema is different, we will throw an error.

language_options: Option<LanguageOptions>

The language specific settings for the translation task.

Implementations

impl TranslationTaskDetails

pub fn file_encoding(&self) -> FileEncoding

Returns the enum value of file_encoding, or the default if the field is set to an invalid enum value.

pub fn set_file_encoding(&mut self, value: FileEncoding)

Sets file_encoding to the provided enum value.

pub fn get_special_token_map(&self, key: &str) -> Option<TokenType>

Returns the enum value for the corresponding key in special_token_map, or None if the entry does not exist or it is not a valid enum value.

pub fn insert_special_token_map( &mut self, key: String, value: TokenType ) -> Option<TokenType>

Inserts a key value pair into special_token_map.