Module google_api_proto::google::cloud::bigquery::storage::v1alpha

Modules

list_metastore_partitions_response

Nested message and enum types in ListMetastorePartitionsResponse.

metastore_partition_service_client

Generated client implementations.

Structs

BatchCreateMetastorePartitionsRequest

Request message for BatchCreateMetastorePartitions.

BatchCreateMetastorePartitionsResponse

Response message for BatchCreateMetastorePartitions.

BatchDeleteMetastorePartitionsRequest

Request message for BatchDeleteMetastorePartitions. The MetastorePartition is uniquely identified by values, which is an ordered list. Hence, there is no separate name or partition id field.

BatchSizeTooLargeError

Structured custom error message for batch size too large error. The error can be attached as error details in the returned rpc Status for more structured error handling in the client.

BatchUpdateMetastorePartitionsRequest

Request message for BatchUpdateMetastorePartitions.

BatchUpdateMetastorePartitionsResponse

Response message for BatchUpdateMetastorePartitions.

CreateMetastorePartitionRequest

Request message for CreateMetastorePartition. The MetastorePartition is uniquely identified by values, which is an ordered list. Hence, there is no separate name or partition id field.

FieldSchema

Schema description of a metastore partition column.

ListMetastorePartitionsRequest

Request message for ListMetastorePartitions.

ListMetastorePartitionsResponse

Response message for ListMetastorePartitions.

MetastorePartition

Information about a Hive partition.

MetastorePartitionList

List of metastore partitions.

MetastorePartitionValues

Represents the values of a metastore partition.

ReadStream

Information about a single stream that is used to read partitions.

SerDeInfo

Serializer and deserializer information.

StorageDescriptor

Contains information about the physical storage of the data in the metastore partition.

StreamList

List of streams.

StreamMetastorePartitionsRequest

The top-level message sent by the client to the Partitions.StreamMetastorePartitions method. Follows the default gRPC streaming maximum size of 4 MB.

StreamMetastorePartitionsResponse

This is the response message sent by the server to the client for the Partitions.StreamMetastorePartitions method when the commit is successful. Server will close the stream after sending this message.

UpdateMetastorePartitionRequest

Request message for UpdateMetastorePartition.