Module google_api_proto::google::cloud::bigquery::v2

Modules

bi_engine_reason

Nested message and enum types in BiEngineReason.

bi_engine_statistics

Nested message and enum types in BiEngineStatistics.

big_lake_configuration

Nested message and enum types in BigLakeConfiguration.

dataset

Nested message and enum types in Dataset.

dataset_access_entry

Nested message and enum types in DatasetAccessEntry.

dataset_service_client

Generated client implementations.

explain_query_stage

Nested message and enum types in ExplainQueryStage.

external_data_configuration

Nested message and enum types in ExternalDataConfiguration.

foreign_type_info

Nested message and enum types in ForeignTypeInfo.

get_dataset_request

Nested message and enum types in GetDatasetRequest.

get_table_request

Nested message and enum types in GetTableRequest.

index_unused_reason

Nested message and enum types in IndexUnusedReason.

job_configuration_extract

Nested message and enum types in JobConfigurationExtract.

job_configuration_load

Nested message and enum types in JobConfigurationLoad.

job_configuration_table_copy

Nested message and enum types in JobConfigurationTableCopy.

job_creation_reason

Nested message and enum types in JobCreationReason.

job_service_client

Generated client implementations.

job_statistics

Nested message and enum types in JobStatistics.

join_restriction_policy

Nested message and enum types in JoinRestrictionPolicy.

linked_dataset_metadata

Nested message and enum types in LinkedDatasetMetadata.

list_jobs_request

Nested message and enum types in ListJobsRequest.

materialized_view

Nested message and enum types in MaterializedView.

ml_statistics

Nested message and enum types in MlStatistics.

model

Nested message and enum types in Model.

model_service_client

Generated client implementations.

partition_skew

Nested message and enum types in PartitionSkew.

privacy_policy

Nested message and enum types in PrivacyPolicy.

project_service_client

Generated client implementations.

query_request

Nested message and enum types in QueryRequest.

range_partitioning

Nested message and enum types in RangePartitioning.

remote_model_info

Nested message and enum types in RemoteModelInfo.

restriction_config

Nested message and enum types in RestrictionConfig.

routine

Nested message and enum types in Routine.

routine_service_client

Generated client implementations.

row_access_policy_service_client

Generated client implementations.

script_options

Nested message and enum types in ScriptOptions.

script_statistics

Nested message and enum types in ScriptStatistics.

search_statistics

Nested message and enum types in SearchStatistics.

spark_statistics

Nested message and enum types in SparkStatistics.

standard_sql_data_type

Nested message and enum types in StandardSqlDataType.

table_field_schema

Nested message and enum types in TableFieldSchema.

table_metadata_cache_usage

Nested message and enum types in TableMetadataCacheUsage.

table_replication_info

Nested message and enum types in TableReplicationInfo.

table_service_client

Generated client implementations.

vector_search_statistics

Nested message and enum types in VectorSearchStatistics.

Structs

Access

An object that defines dataset access for an entity.

AggregationThresholdPolicy

Represents privacy policy associated with “aggregation threshold” method.

AvroOptions

Options for external data sources.

BiEngineReason

Reason why BI Engine didn’t accelerate the query (or sub-query).

BiEngineStatistics

Statistics for a BI Engine specific query. Populated as part of JobStatistics2

BigLakeConfiguration

Configuration for BigLake managed tables.

BigtableColumn

Information related to a Bigtable column.

BigtableColumnFamily

Information related to a Bigtable column family.

BigtableOptions

Options specific to Google Cloud Bigtable data sources.

CancelJobRequest

Describes format of a jobs cancellation request.

CloneDefinition

Information about base table and clone time of a table clone.

Clustering

Configures table clustering.

ColumnReference

The pair of the foreign key column and primary key column.

ConnectionProperty

A connection-level property to customize query behavior. Under JDBC, these correspond directly to connection properties passed to the DriverManager. Under ODBC, these correspond to properties in the connection string.

CopyJobStatistics

Statistics for a copy job.

CsvOptions

Information related to a CSV data source.

DataFormatOptions

Options for data format adjustments.

DataMaskingStatistics

Statistics for data-masking.

DataPolicyOption

Data policy option proto, it currently supports name only, will support precedence later.

Dataset

Represents a BigQuery dataset.

DatasetAccessEntry

Grants all resources of particular types in a particular dataset read access to the current dataset.

DatasetList

Response format for a page of results when listing datasets.

DatasetReference

Identifier for a dataset.

DeleteDatasetRequest

Request format for deleting a dataset.

DeleteJobRequest

Describes the format of a jobs deletion request.

DeleteModelRequest

Request format for deleting BigQuery ML models.

DeleteRoutineRequest

Describes the format for deleting a routine.

DeleteTableRequest

Request format for deleting a table.

DestinationTableProperties

Properties for the destination table.

DifferentialPrivacyPolicy

Represents privacy policy associated with “differential privacy” method.

DmlStats

Detailed statistics for DML statements

EncryptionConfiguration

Configuration for Cloud KMS encryption settings.

ErrorProto

Error details.

ExplainQueryStage

A single stage of query execution.

ExplainQueryStep

An operation within a stage.

ExportDataStatistics

Statistics for the EXPORT DATA statement as part of Query Job. EXTRACT JOB statistics are populated in JobStatistics4.

ExternalCatalogDatasetOptions

Options defining open source compatible datasets living in the BigQuery catalog. Contains metadata of open source database, schema or namespace represented by the current dataset.

ExternalCatalogTableOptions

Metadata about open source compatible table. The fields contained in these options correspond to hive metastore’s table level properties.

ExternalDataConfiguration

ExternalDatasetReference

Configures the access a dataset defined in an external metadata storage.

ExternalServiceCost

The external service cost is a portion of the total cost, these costs are not additive with total_bytes_billed. Moreover, this field only track external service costs that will show up as BigQuery costs (e.g. training BigQuery ML job with google cloud CAIP or Automl Tables services), not other costs which may be accrued by running the query (e.g. reading from Bigtable or Cloud Storage). The external service costs with different billing sku (e.g. CAIP job is charged based on VM usage) are converted to BigQuery billed_bytes and slot_ms with equivalent amount of US dollars. Services may not directly correlate to these metrics, but these are the equivalents for billing purposes. Output only.

ForeignKey

Represents a foreign key constraint on a table’s columns.

ForeignTypeInfo

Metadata about the foreign data type definition such as the system in which the type is defined.

ForeignViewDefinition

A view can be represented in multiple ways. Each representation has its own dialect. This message stores the metadata required for these representations.

GcpTag

A global tag managed by Resource Manager. https://cloud.google.com/iam/docs/tags-access-control#definitions

GetDatasetRequest

Request format for getting information about a dataset.

GetJobRequest

Describes format of a jobs get request.

GetModelRequest

Request format for getting information about a BigQuery ML model.

GetQueryResultsRequest

Request object of GetQueryResults.

GetQueryResultsResponse

Response object of GetQueryResults.

GetRoutineRequest

Describes the format for getting information about a routine.

GetServiceAccountRequest

Request object of GetServiceAccount

GetServiceAccountResponse

Response object of GetServiceAccount

GetTableRequest

Request format for getting table metadata.

GoogleSheetsOptions

Options specific to Google Sheets data sources.

HighCardinalityJoin

High cardinality join detailed information.

HivePartitioningOptions

Options for configuring hive partitioning detect.

IndexUnusedReason

Reason about why no search index was used in the search query (or sub-query).

InputDataChange

Details about the input data change insight.

InsertDatasetRequest

Request format for inserting a dataset.

InsertJobRequest

Describes format of a job insertion request.

InsertRoutineRequest

Describes the format for inserting a routine.

InsertTableRequest

Request format for inserting table metadata.

Job

JobCancelResponse

Describes format of a jobs cancellation response.

JobConfiguration

JobConfigurationExtract

JobConfigurationExtract configures a job that exports data from a BigQuery table into Google Cloud Storage.

JobConfigurationLoad

JobConfigurationLoad contains the configuration properties for loading data into a destination table.

JobConfigurationQuery

JobConfigurationQuery configures a BigQuery query job.

JobConfigurationTableCopy

JobConfigurationTableCopy configures a job that copies data from one table to another. For more information on copying tables, see Copy a table.

JobCreationReason

Reason about why a Job was created from a jobs.query method when used with JOB_CREATION_OPTIONAL Job creation mode.

JobList

JobList is the response format for a jobs.list call.

JobReference

A job reference is a fully qualified identifier for referring to a job.

JobStatistics

Statistics for a single job execution.

JobStatistics2

Statistics for a query job.

JobStatistics3

Statistics for a load job.

JobStatistics4

Statistics for an extract job.

JobStatus

JoinRestrictionPolicy

Represents privacy policy associated with “join restrictions”. Join restriction gives data providers the ability to enforce joins on the ‘join_allowed_columns’ when data is queried from a privacy protected view.

JsonOptions

Json Options for load and make external tables.

LinkedDatasetMetadata

Metadata about the Linked Dataset.

LinkedDatasetSource

A dataset source type which refers to another BigQuery dataset.

ListDatasetsRequest

ListFormatDataset

A dataset resource with only a subset of fields, to be returned in a list of datasets.

ListFormatJob

ListFormatJob is a partial projection of job information returned as part of a jobs.list response.

ListFormatTable

ListFormatView

Information about a logical view.

ListJobsRequest

Describes the format of the list jobs request.

ListModelsRequest

Request format for listing BigQuery ML models.

ListModelsResponse

Response format for a single page when listing BigQuery ML models.

ListRoutinesRequest

Describes the format for listing routines.

ListRoutinesResponse

Describes the format of a single result page when listing routines.

ListRowAccessPoliciesRequest

Request message for the ListRowAccessPolicies method.

ListRowAccessPoliciesResponse

Response message for the ListRowAccessPolicies method.

ListTablesRequest

Request format for enumerating tables.

LoadQueryStatistics

Statistics for a LOAD query.

LocationMetadata

BigQuery-specific metadata about a location. This will be set on google.cloud.location.Location.metadata in Cloud Location API responses.

MaterializedView

A materialized view considered for a query job.

MaterializedViewDefinition

Definition and configuration of a materialized view.

MaterializedViewStatistics

Statistics of materialized views considered in a query job.

MaterializedViewStatus

Status of a materialized view. The last refresh timestamp status is omitted here, but is present in the MaterializedViewDefinition message.

MetadataCacheStatistics

Statistics for metadata caching in BigLake tables.

MlStatistics

Job statistics specific to a BigQuery ML training job.

Model

ModelReference

Id path of a model.

ParquetOptions

Parquet Options for load and make external tables.

PartitionSkew

Partition skew detailed information.

PartitionedColumn

The partitioning column information.

PartitioningDefinition

The partitioning information, which includes managed table, external table and metastore partitioned table partition information.

PatchModelRequest

PatchRoutineRequest

Describes the format for the partial update (patch) of a routine.

PerformanceInsights

Performance insights for the job.

PostQueryRequest

Request format for the query request.

PrimaryKey

Represents the primary key constraint on a table’s columns.

PrivacyPolicy

Represents privacy policy that contains the privacy requirements specified by the data owner. Currently, this is only supported on views.

QueryInfo

Query optimization information for a QUERY job.

QueryParameter

A parameter given to a query.

QueryParameterStructType

The type of a struct parameter.

QueryParameterType

The type of a query parameter.

QueryParameterValue

The value of a query parameter.

QueryRequest

Describes the format of the jobs.query request.

QueryResponse

QueryTimelineSample

Summary of the state of query execution at a given time.

RangePartitioning

RangeValue

Represents the value of a range.

RemoteModelInfo

Remote Model Info

RestrictionConfig

Routine

A user-defined function or a stored procedure.

RoutineReference

Id path of a routine.

RowAccessPolicy

Represents access on a subset of rows on the specified table, defined by its filter predicate. Access to the subset of rows is controlled by its IAM policy.

RowAccessPolicyReference

Id path of a row access policy.

RowLevelSecurityStatistics

Statistics for row-level security.

ScriptOptions

Options related to script execution.

ScriptStatistics

Job statistics specific to the child job of a script.

SearchStatistics

Statistics for a search query. Populated as part of JobStatistics2.

SerDeInfo

Serializer and deserializer information.

SessionInfo

[Preview] Information related to sessions.

SnapshotDefinition

Information about base table and snapshot time of the snapshot.

SparkOptions

Options for a user-defined Spark routine.

SparkStatistics

Statistics for a BigSpark query. Populated as part of JobStatistics2

StagePerformanceChangeInsight

Performance insights compared to the previous executions for a specific stage.

StagePerformanceStandaloneInsight

Standalone performance insights for a specific stage.

StandardSqlDataType

The data type of a variable such as a function argument. Examples include:

StandardSqlField

A field or a column.

StandardSqlStructType

The representation of a SQL STRUCT type.

StandardSqlTableType

A table type

StorageDescriptor

Contains information about how a table’s data is stored and accessed by open source query engines.

Streamingbuffer

SystemVariables

System variables given to a query.

Table

TableConstraints

The TableConstraints defines the primary key and foreign key.

TableFieldSchema

A field in TableSchema

TableList

Partial projection of the metadata for a given table in a list response.

TableMetadataCacheUsage

Table level detail on the usage of metadata caching. Only set for Metadata caching eligible tables referenced in the query.

TableReference

TableReplicationInfo

Replication info of a table created using AS REPLICA DDL like: CREATE MATERIALIZED VIEW mv1 AS REPLICA OF src_mv

TableSchema

Schema of a table

TimePartitioning

TransformColumn

Information about a single transform column.

UndeleteDatasetRequest

Request format for undeleting a dataset.

UpdateOrPatchDatasetRequest

Message for updating or patching a dataset.

UpdateOrPatchTableRequest

UpdateRoutineRequest

Describes the format for updating a routine.

UserDefinedFunctionResource

This is used for defining User Defined Function (UDF) resources only when using legacy SQL. Users of GoogleSQL should leverage either DDL (e.g. CREATE [TEMPORARY] FUNCTION … ) or the Routines API to define UDF resources.

VectorSearchStatistics

Statistics for a vector search query. Populated as part of JobStatistics2.

ViewDefinition

Describes the definition of a logical view.

Enums

DecimalTargetType

The data types that could be used as a target type when converting decimal values.

FileSetSpecType

This enum defines how to interpret source URIs for load jobs and external tables.

JsonExtension

Used to indicate that a JSON variant, rather than normal JSON, is being used as the source_format. This should only be used in combination with the JSON source format.

MapTargetType

Indicates the map target type. Only applies to parquet maps.

ReservationEdition

The type of editions. Different features and behaviors are provided to different editions Capacity commitments and reservations are linked to editions.