Module google_api_proto::google::cloud::bigquery::v2::model

Nested message and enum types in Model.

Modules

arima_forecasting_metrics

Nested message and enum types in ArimaForecastingMetrics.

binary_classification_metrics

Nested message and enum types in BinaryClassificationMetrics.

boosted_tree_option_enums

Nested message and enum types in BoostedTreeOptionEnums.

category_encoding_method

Nested message and enum types in CategoryEncodingMethod.

clustering_metrics

Nested message and enum types in ClusteringMetrics.

double_hparam_search_space

Nested message and enum types in DoubleHparamSearchSpace.

evaluation_metrics

Nested message and enum types in EvaluationMetrics.

global_explanation

Nested message and enum types in GlobalExplanation.

hparam_tuning_enums

Nested message and enum types in HparamTuningEnums.

hparam_tuning_trial

Nested message and enum types in HparamTuningTrial.

int_array_hparam_search_space

Nested message and enum types in IntArrayHparamSearchSpace.

int_hparam_search_space

Nested message and enum types in IntHparamSearchSpace.

kmeans_enums

Nested message and enum types in KmeansEnums.

model_registry_option_enums

Nested message and enum types in ModelRegistryOptionEnums.

multi_class_classification_metrics

Nested message and enum types in MultiClassClassificationMetrics.

pca_solver_option_enums

Nested message and enum types in PcaSolverOptionEnums.

seasonal_period

Nested message and enum types in SeasonalPeriod.

training_run

Nested message and enum types in TrainingRun.

Structs

AggregateClassificationMetrics

Aggregate metrics for classification/classifier models. For multi-class models, the metrics are either macro-averaged or micro-averaged. When macro-averaged, the metrics are calculated for each label and then an unweighted average is taken of those values. When micro-averaged, the metric is calculated globally by counting the total number of correctly predicted rows.

ArimaFittingMetrics

ARIMA model fitting metrics.

ArimaForecastingMetrics

Model evaluation metrics for ARIMA forecasting models.

ArimaOrder

Arima order, can be used for both non-seasonal and seasonal parts.

BinaryClassificationMetrics

Evaluation metrics for binary classification/classifier models.

BoostedTreeOptionEnums

Enums for XGBoost model type.

CategoryEncodingMethod

Encoding methods for categorical features.

ClusteringMetrics

Evaluation metrics for clustering models.

DataSplitResult

Data split result. This contains references to the training and evaluation data tables that were used to train the model.

DimensionalityReductionMetrics

Model evaluation metrics for dimensionality reduction models.

DoubleHparamSearchSpace

Search space for a double hyperparameter.

EvaluationMetrics

Evaluation metrics of a model. These are either computed on all training data or just the eval data based on whether eval data was used during training. These are not present for imported models.

GlobalExplanation

Global explanations containing the top most important features after training.

HparamSearchSpaces

Hyperparameter search spaces. These should be a subset of training_options.

HparamTuningEnums

Enums for hyperparameter tuning.

HparamTuningTrial

Training info of a trial in hyperparameter tuning models.

IntArrayHparamSearchSpace

Search space for int array.

IntHparamSearchSpace

Search space for an int hyperparameter.

KmeansEnums

Enums for kmeans model type.

ModelRegistryOptionEnums

Model registry options.

MultiClassClassificationMetrics

Evaluation metrics for multi-class classification/classifier models.

PcaSolverOptionEnums

PCA solver options.

RankingMetrics

Evaluation metrics used by weighted-ALS models specified by feedback_type=implicit.

RegressionMetrics

Evaluation metrics for regression and explicit feedback type matrix factorization models.

SeasonalPeriod

Enums for seasonal period.

StringHparamSearchSpace

Search space for string and enum.

TrainingRun

Information about a single training query run for the model.

Enums

ColorSpace

Enums for color space, used for processing images in Object Table. See more details at https://www.tensorflow.org/io/tutorials/colorspace.

DataFrequency

Type of supported data frequency for time series forecasting models.

DataSplitMethod

Indicates the method to split input data into multiple tables.

DistanceType

Distance metric used to compute the distance between two points.

FeedbackType

Indicates the training algorithm to use for matrix factorization models.

HolidayRegion

Type of supported holiday regions for time series forecasting models.

LearnRateStrategy

Indicates the learning rate optimization strategy to use.

LossType

Loss metric to evaluate model training performance.

ModelType

Indicates the type of the Model.

OptimizationStrategy

Indicates the optimization strategy used for training.