Enum google_api_proto::google::cloud::bigquery::v2::model::HolidayRegion

#[repr(i32)]
pub enum HolidayRegion {

    Unspecified = 0,
    Global = 1,
    Na = 2,
    Japac = 3,
    Emea = 4,
    Lac = 5,
    Ae = 6,
    Ar = 7,
    At = 8,
    Au = 9,
    Be = 10,
    Br = 11,
    Ca = 12,
    Ch = 13,
    Cl = 14,
    Cn = 15,
    Co = 16,
    Cs = 17,
    Cz = 18,
    De = 19,
    Dk = 20,
    Dz = 21,
    Ec = 22,
    Ee = 23,
    Eg = 24,
    Es = 25,
    Fi = 26,
    Fr = 27,
    Gb = 28,
    Gr = 29,
    Hk = 30,
    Hu = 31,
    Id = 32,
    Ie = 33,
    Il = 34,
    In = 35,
    Ir = 36,
    It = 37,
    Jp = 38,
    Kr = 39,
    Lv = 40,
    Ma = 41,
    Mx = 42,
    My = 43,
    Ng = 44,
    Nl = 45,
    No = 46,
    Nz = 47,
    Pe = 48,
    Ph = 49,
    Pk = 50,
    Pl = 51,
    Pt = 52,
    Ro = 53,
    Rs = 54,
    Ru = 55,
    Sa = 56,
    Se = 57,
    Sg = 58,
    Si = 59,
    Sk = 60,
    Th = 61,
    Tr = 62,
    Tw = 63,
    Ua = 64,
    Us = 65,
    Ve = 66,
    Vn = 67,
    Za = 68,

}

Type of supported holiday regions for time series forecasting models.

Variants

Unspecified = 0

Holiday region unspecified.

Global = 1

Global.

Na = 2

North America.

Japac = 3

Japan and Asia Pacific: Korea, Greater China, India, Australia, and New Zealand.

Emea = 4

Europe, the Middle East and Africa.

Lac = 5

Latin America and the Caribbean.

Ae = 6

United Arab Emirates

Ar = 7

Argentina

At = 8

Austria

Au = 9

Australia

Be = 10

Belgium

Br = 11

Brazil

Ca = 12

Canada

Ch = 13

Switzerland

Cl = 14

Chile

Cn = 15

China

Co = 16

Colombia

Cs = 17

Czechoslovakia

Cz = 18

Czech Republic

De = 19

Germany

Dk = 20

Denmark

Dz = 21

Algeria

Ec = 22

Ecuador

Ee = 23

Estonia

Eg = 24

Egypt

Es = 25

Spain

Fi = 26

Finland

Fr = 27

France

Gb = 28

Great Britain (United Kingdom)

Gr = 29

Greece

Hk = 30

Hong Kong

Hu = 31

Hungary

Id = 32

Indonesia

Ie = 33

Ireland

Il = 34

Israel

In = 35

India

Ir = 36

Iran

It = 37

Italy

Jp = 38

Japan

Kr = 39

Korea (South)

Lv = 40

Latvia

Ma = 41

Morocco

Mx = 42

Mexico

My = 43

Malaysia

Ng = 44

Nigeria

Nl = 45

Netherlands

No = 46

Norway

Nz = 47

New Zealand

Pe = 48

Peru

Ph = 49

Philippines

Pk = 50

Pakistan

Pl = 51

Poland

Pt = 52

Portugal

Ro = 53

Romania

Rs = 54

Serbia

Ru = 55

Russian Federation

Sa = 56

Saudi Arabia

Se = 57

Sweden

Sg = 58

Singapore

Si = 59

Slovenia

Sk = 60

Slovakia

Th = 61

Thailand

Tr = 62

Turkey

Tw = 63

Taiwan

Ua = 64

Ukraine

Us = 65

United States

Ve = 66

Venezuela

Vn = 67

Viet Nam

Za = 68

South Africa

Implementations

impl HolidayRegion

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of HolidayRegion.

pub fn from_i32(value: i32) -> Option<HolidayRegion>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a HolidayRegion, or None if value is not a valid variant.

impl HolidayRegion

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.