Struct google_api_proto::google::cloud::bigquery::v2::CsvOptions

pub struct CsvOptions {
    pub field_delimiter: String,
    pub skip_leading_rows: Option<i64>,
    pub quote: Option<String>,
    pub allow_quoted_newlines: Option<bool>,
    pub allow_jagged_rows: Option<bool>,
    pub encoding: String,
    pub preserve_ascii_control_characters: Option<bool>,
    pub null_marker: Option<String>,
}

Information related to a CSV data source.

Fields
field_delimiter: String

Optional. The separator character for fields in a CSV file. The separator is interpreted as a single byte. For files encoded in ISO-8859-1, any single character can be used as a separator. For files encoded in UTF-8, characters represented in decimal range 1-127 (U+0001-U+007F) can be used without any modification. UTF-8 characters encoded with multiple bytes (i.e. U+0080 and above) will have only the first byte used for separating fields. The remaining bytes will be treated as a part of the field. BigQuery also supports the escape sequence “\t” (U+0009) to specify a tab separator. The default value is comma (“,”, U+002C).

skip_leading_rows: Option<i64>

Optional. The number of rows at the top of a CSV file that BigQuery will skip when reading the data. The default value is 0. This property is useful if you have header rows in the file that should be skipped. When autodetect is on, the behavior is the following:

skipLeadingRows unspecified - Autodetect tries to detect headers in the first row. If they are not detected, the row is read as data. Otherwise data is read starting from the second row.

skipLeadingRows is 0 - Instructs autodetect that there are no headers and data should be read starting from the first row.

skipLeadingRows = N > 0 - Autodetect skips N-1 rows and tries to detect headers in row N. If headers are not detected, row N is just skipped. Otherwise row N is used to extract column names for the detected schema.

quote: Option<String>

Optional. The value that is used to quote data sections in a CSV file. BigQuery converts the string to ISO-8859-1 encoding, and then uses the first byte of the encoded string to split the data in its raw, binary state. The default value is a double-quote (“). If your data does not contain quoted sections, set the property value to an empty string. If your data contains quoted newline characters, you must also set the allowQuotedNewlines property to true. To include the specific quote character within a quoted value, precede it with an additional matching quote character. For example, if you want to escape the default character ’ “ ’, use ’ “” ’.

allow_quoted_newlines: Option<bool>

Optional. Indicates if BigQuery should allow quoted data sections that contain newline characters in a CSV file. The default value is false.

allow_jagged_rows: Option<bool>

Optional. Indicates if BigQuery should accept rows that are missing trailing optional columns. If true, BigQuery treats missing trailing columns as null values. If false, records with missing trailing columns are treated as bad records, and if there are too many bad records, an invalid error is returned in the job result. The default value is false.

encoding: String

Optional. The character encoding of the data. The supported values are UTF-8, ISO-8859-1, UTF-16BE, UTF-16LE, UTF-32BE, and UTF-32LE. The default value is UTF-8. BigQuery decodes the data after the raw, binary data has been split using the values of the quote and fieldDelimiter properties.

preserve_ascii_control_characters: Option<bool>

Optional. Indicates if the embedded ASCII control characters (the first 32 characters in the ASCII-table, from ‘\x00’ to ‘\x1F’) are preserved.

null_marker: Option<String>

Optional. Specifies a string that represents a null value in a CSV file. For example, if you specify “\N”, BigQuery interprets “\N” as a null value when querying a CSV file. The default value is the empty string. If you set this property to a custom value, BigQuery throws an error if an empty string is present for all data types except for STRING and BYTE. For STRING and BYTE columns, BigQuery interprets the empty string as an empty value.