Struct google_api_proto::google::cloud::bigquery::v2::JobStatistics

pub struct JobStatistics {

    pub creation_time: i64,
    pub start_time: i64,
    pub end_time: i64,
    pub total_bytes_processed: Option<i64>,
    pub completion_ratio: Option<f64>,
    pub quota_deferments: Vec<String>,
    pub query: Option<JobStatistics2>,
    pub load: Option<JobStatistics3>,
    pub extract: Option<JobStatistics4>,
    pub copy: Option<CopyJobStatistics>,
    pub total_slot_ms: Option<i64>,
    pub reservation_id: String,
    pub num_child_jobs: i64,
    pub parent_job_id: String,
    pub script_statistics: Option<ScriptStatistics>,
    pub row_level_security_statistics: Option<RowLevelSecurityStatistics>,
    pub data_masking_statistics: Option<DataMaskingStatistics>,
    pub transaction_info: Option<TransactionInfo>,
    pub session_info: Option<SessionInfo>,
    pub final_execution_duration_ms: i64,
    pub edition: i32,

}

Statistics for a single job execution.

Fields
creation_time: i64

Output only. Creation time of this job, in milliseconds since the epoch. This field will be present on all jobs.

start_time: i64

Output only. Start time of this job, in milliseconds since the epoch. This field will be present when the job transitions from the PENDING state to either RUNNING or DONE.

end_time: i64

Output only. End time of this job, in milliseconds since the epoch. This field will be present whenever a job is in the DONE state.

total_bytes_processed: Option<i64>

Output only. Total bytes processed for the job.

completion_ratio: Option<f64>

Output only. [TrustedTester] Job progress (0.0 -> 1.0) for LOAD and EXTRACT jobs.

quota_deferments: Vec<String>

Output only. Quotas which delayed this job’s start time.

query: Option<JobStatistics2>

Output only. Statistics for a query job.

load: Option<JobStatistics3>

Output only. Statistics for a load job.

extract: Option<JobStatistics4>

Output only. Statistics for an extract job.

copy: Option<CopyJobStatistics>

Output only. Statistics for a copy job.

total_slot_ms: Option<i64>

Output only. Slot-milliseconds for the job.

reservation_id: String

Output only. Name of the primary reservation assigned to this job. Note that this could be different than reservations reported in the reservation usage field if parent reservations were used to execute this job.

num_child_jobs: i64

Output only. Number of child jobs executed.

parent_job_id: String

Output only. If this is a child job, specifies the job ID of the parent.

script_statistics: Option<ScriptStatistics>

Output only. If this a child job of a script, specifies information about the context of this job within the script.

row_level_security_statistics: Option<RowLevelSecurityStatistics>

Output only. Statistics for row-level security. Present only for query and extract jobs.

data_masking_statistics: Option<DataMaskingStatistics>

Output only. Statistics for data-masking. Present only for query and extract jobs.

transaction_info: Option<TransactionInfo>

Output only. [Alpha] Information of the multi-statement transaction if this job is part of one.

This property is only expected on a child job or a job that is in a session. A script parent job is not part of the transaction started in the script.

session_info: Option<SessionInfo>

Output only. Information of the session if this job is part of one.

final_execution_duration_ms: i64

Output only. The duration in milliseconds of the execution of the final attempt of this job, as BigQuery may internally re-attempt to execute the job.

edition: i32

Output only. Name of edition corresponding to the reservation for this job at the time of this update.

Implementations

impl JobStatistics

pub fn edition(&self) -> ReservationEdition

Returns the enum value of edition, or the default if the field is set to an invalid enum value.

pub fn set_edition(&mut self, value: ReservationEdition)

Sets edition to the provided enum value.