Struct google_api_proto::google::cloud::bigquery::v2::JobStatistics2

pub struct JobStatistics2 {

    pub query_plan: Vec<ExplainQueryStage>,
    pub estimated_bytes_processed: Option<i64>,
    pub timeline: Vec<QueryTimelineSample>,
    pub total_partitions_processed: Option<i64>,
    pub total_bytes_processed: Option<i64>,
    pub total_bytes_processed_accuracy: String,
    pub total_bytes_billed: Option<i64>,
    pub billing_tier: Option<i32>,
    pub total_slot_ms: Option<i64>,
    pub cache_hit: Option<bool>,
    pub referenced_tables: Vec<TableReference>,
    pub referenced_routines: Vec<RoutineReference>,
    pub schema: Option<TableSchema>,
    pub num_dml_affected_rows: Option<i64>,
    pub dml_stats: Option<DmlStats>,
    pub undeclared_query_parameters: Vec<QueryParameter>,
    pub statement_type: String,
    pub ddl_operation_performed: String,
    pub ddl_target_table: Option<TableReference>,
    pub ddl_destination_table: Option<TableReference>,
    pub ddl_target_row_access_policy: Option<RowAccessPolicyReference>,
    pub ddl_affected_row_access_policy_count: Option<i64>,
    pub ddl_target_routine: Option<RoutineReference>,
    pub ddl_target_dataset: Option<DatasetReference>,
    pub ml_statistics: Option<MlStatistics>,
    pub export_data_statistics: Option<ExportDataStatistics>,
    pub external_service_costs: Vec<ExternalServiceCost>,
    pub bi_engine_statistics: Option<BiEngineStatistics>,
    pub load_query_statistics: Option<LoadQueryStatistics>,
    pub dcl_target_table: Option<TableReference>,
    pub dcl_target_view: Option<TableReference>,
    pub dcl_target_dataset: Option<DatasetReference>,
    pub search_statistics: Option<SearchStatistics>,
    pub vector_search_statistics: Option<VectorSearchStatistics>,
    pub performance_insights: Option<PerformanceInsights>,
    pub query_info: Option<QueryInfo>,
    pub spark_statistics: Option<SparkStatistics>,
    pub transferred_bytes: Option<i64>,
    pub materialized_view_statistics: Option<MaterializedViewStatistics>,
    pub metadata_cache_statistics: Option<MetadataCacheStatistics>,

}

Statistics for a query job.

Fields
query_plan: Vec<ExplainQueryStage>

Output only. Describes execution plan for the query.

estimated_bytes_processed: Option<i64>

Output only. The original estimate of bytes processed for the job.

timeline: Vec<QueryTimelineSample>

Output only. Describes a timeline of job execution.

total_partitions_processed: Option<i64>

Output only. Total number of partitions processed from all partitioned tables referenced in the job.

total_bytes_processed: Option<i64>

Output only. Total bytes processed for the job.

total_bytes_processed_accuracy: String

Output only. For dry-run jobs, totalBytesProcessed is an estimate and this field specifies the accuracy of the estimate. Possible values can be: UNKNOWN: accuracy of the estimate is unknown. PRECISE: estimate is precise. LOWER_BOUND: estimate is lower bound of what the query would cost. UPPER_BOUND: estimate is upper bound of what the query would cost.

total_bytes_billed: Option<i64>

Output only. If the project is configured to use on-demand pricing, then this field contains the total bytes billed for the job. If the project is configured to use flat-rate pricing, then you are not billed for bytes and this field is informational only.

billing_tier: Option<i32>

Output only. Billing tier for the job. This is a BigQuery-specific concept which is not related to the Google Cloud notion of “free tier”. The value here is a measure of the query’s resource consumption relative to the amount of data scanned. For on-demand queries, the limit is 100, and all queries within this limit are billed at the standard on-demand rates. On-demand queries that exceed this limit will fail with a billingTierLimitExceeded error.

total_slot_ms: Option<i64>

Output only. Slot-milliseconds for the job.

cache_hit: Option<bool>

Output only. Whether the query result was fetched from the query cache.

referenced_tables: Vec<TableReference>

Output only. Referenced tables for the job. Queries that reference more than 50 tables will not have a complete list.

referenced_routines: Vec<RoutineReference>

Output only. Referenced routines for the job.

schema: Option<TableSchema>

Output only. The schema of the results. Present only for successful dry run of non-legacy SQL queries.

num_dml_affected_rows: Option<i64>

Output only. The number of rows affected by a DML statement. Present only for DML statements INSERT, UPDATE or DELETE.

dml_stats: Option<DmlStats>

Output only. Detailed statistics for DML statements INSERT, UPDATE, DELETE, MERGE or TRUNCATE.

undeclared_query_parameters: Vec<QueryParameter>

Output only. GoogleSQL only: list of undeclared query parameters detected during a dry run validation.

statement_type: String

Output only. The type of query statement, if valid. Possible values:

SELECT: SELECT statement.

ASSERT: ASSERT statement.

INSERT: INSERT statement.

UPDATE: UPDATE statement.

DELETE: DELETE statement.

MERGE: MERGE statement.

CREATE_TABLE: CREATE TABLE statement, without AS SELECT.

CREATE_TABLE_AS_SELECT: CREATE TABLE AS SELECT statement.

CREATE_VIEW: CREATE VIEW statement.

CREATE_MODEL: CREATE MODEL statement.

CREATE_MATERIALIZED_VIEW: CREATE MATERIALIZED VIEW statement.

CREATE_FUNCTION: CREATE FUNCTION statement.

CREATE_TABLE_FUNCTION: CREATE TABLE FUNCTION statement.

CREATE_PROCEDURE: CREATE PROCEDURE statement.

CREATE_ROW_ACCESS_POLICY: CREATE ROW ACCESS POLICY statement.

CREATE_SCHEMA: CREATE SCHEMA statement.

CREATE_SNAPSHOT_TABLE: CREATE SNAPSHOT TABLE statement.

CREATE_SEARCH_INDEX: CREATE SEARCH INDEX statement.

DROP_TABLE: DROP TABLE statement.

DROP_EXTERNAL_TABLE: DROP EXTERNAL TABLE statement.

DROP_VIEW: DROP VIEW statement.

DROP_MODEL: DROP MODEL statement.

DROP_MATERIALIZED_VIEW: DROP MATERIALIZED VIEW statement.

DROP_FUNCTION : DROP FUNCTION statement.

DROP_TABLE_FUNCTION : DROP TABLE FUNCTION statement.

DROP_PROCEDURE: DROP PROCEDURE statement.

DROP_SEARCH_INDEX: DROP SEARCH INDEX statement.

DROP_SCHEMA: DROP SCHEMA statement.

DROP_SNAPSHOT_TABLE: DROP SNAPSHOT TABLE statement.

DROP_ROW_ACCESS_POLICY: [DROP [ALL\] ROW ACCESS POLICY|POLICIES](https://cloud.google.com/bigquery/docs/reference/standard-sql/data-definition-language#drop_row_access_policy_statement) statement.

ALTER_TABLE: ALTER TABLE statement.

ALTER_VIEW: ALTER VIEW statement.

ALTER_MATERIALIZED_VIEW: ALTER MATERIALIZED VIEW statement.

ALTER_SCHEMA: ALTER SCHEMA statement.

SCRIPT: SCRIPT.

TRUNCATE_TABLE: TRUNCATE TABLE statement.

CREATE_EXTERNAL_TABLE: CREATE EXTERNAL TABLE statement.

EXPORT_DATA: EXPORT DATA statement.

EXPORT_MODEL: EXPORT MODEL statement.

LOAD_DATA: LOAD DATA statement.

CALL: CALL statement.

ddl_operation_performed: String

Output only. The DDL operation performed, possibly dependent on the pre-existence of the DDL target.

ddl_target_table: Option<TableReference>

Output only. The DDL target table. Present only for CREATE/DROP TABLE/VIEW and DROP ALL ROW ACCESS POLICIES queries.

ddl_destination_table: Option<TableReference>

Output only. The table after rename. Present only for ALTER TABLE RENAME TO query.

ddl_target_row_access_policy: Option<RowAccessPolicyReference>

Output only. The DDL target row access policy. Present only for CREATE/DROP ROW ACCESS POLICY queries.

ddl_affected_row_access_policy_count: Option<i64>

Output only. The number of row access policies affected by a DDL statement. Present only for DROP ALL ROW ACCESS POLICIES queries.

ddl_target_routine: Option<RoutineReference>

Output only. [Beta] The DDL target routine. Present only for CREATE/DROP FUNCTION/PROCEDURE queries.

ddl_target_dataset: Option<DatasetReference>

Output only. The DDL target dataset. Present only for CREATE/ALTER/DROP SCHEMA(dataset) queries.

ml_statistics: Option<MlStatistics>

Output only. Statistics of a BigQuery ML training job.

export_data_statistics: Option<ExportDataStatistics>

Output only. Stats for EXPORT DATA statement.

external_service_costs: Vec<ExternalServiceCost>

Output only. Job cost breakdown as bigquery internal cost and external service costs.

bi_engine_statistics: Option<BiEngineStatistics>

Output only. BI Engine specific Statistics.

load_query_statistics: Option<LoadQueryStatistics>

Output only. Statistics for a LOAD query.

dcl_target_table: Option<TableReference>

Output only. Referenced table for DCL statement.

dcl_target_view: Option<TableReference>

Output only. Referenced view for DCL statement.

dcl_target_dataset: Option<DatasetReference>

Output only. Referenced dataset for DCL statement.

search_statistics: Option<SearchStatistics>

Output only. Search query specific statistics.

vector_search_statistics: Option<VectorSearchStatistics>

Output only. Vector Search query specific statistics.

performance_insights: Option<PerformanceInsights>

Output only. Performance insights.

query_info: Option<QueryInfo>

Output only. Query optimization information for a QUERY job.

spark_statistics: Option<SparkStatistics>

Output only. Statistics of a Spark procedure job.

transferred_bytes: Option<i64>

Output only. Total bytes transferred for cross-cloud queries such as Cross Cloud Transfer and CREATE TABLE AS SELECT (CTAS).

materialized_view_statistics: Option<MaterializedViewStatistics>

Output only. Statistics of materialized views of a query job.

metadata_cache_statistics: Option<MetadataCacheStatistics>

Output only. Statistics of metadata cache usage in a query for BigLake tables.