Struct google_api_proto::google::cloud::certificatemanager::v1::certificate_manager_client::CertificateManagerClient

pub struct CertificateManagerClient<T> { /* private fields */ }

API Overview

Certificates Manager API allows customers to see and manage all their TLS certificates.

Certificates Manager API service provides methods to manage certificates, group them into collections, and create serving configuration that can be easily applied to other Cloud resources e.g. Target Proxies.

Data Model

The Certificates Manager service exposes the following resources:

Certificate that describes a single TLS certificate.

CertificateMap that describes a collection of certificates that can be attached to a target resource.

CertificateMapEntry that describes a single configuration entry that consists of a SNI and a group of certificates. It’s a subresource of CertificateMap.

Certificate, CertificateMap and CertificateMapEntry IDs have to fully match the regexp [a-z0-9-]{1,63}. In other words,

only lower case letters, digits, and hyphen are allowed

length of the resource ID has to be in [1,63] range.

Provides methods to manage Cloud Certificate Manager entities.

Implementations

impl<T> CertificateManagerClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> CertificateManagerClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_certificates( &mut self, request: impl IntoRequest<ListCertificatesRequest> ) -> Result<Response<ListCertificatesResponse>, Status>

Lists Certificates in a given project and location.

pub async fn get_certificate( &mut self, request: impl IntoRequest<GetCertificateRequest> ) -> Result<Response<Certificate>, Status>

Gets details of a single Certificate.

pub async fn create_certificate( &mut self, request: impl IntoRequest<CreateCertificateRequest> ) -> Result<Response<Operation>, Status>

Creates a new Certificate in a given project and location.

pub async fn update_certificate( &mut self, request: impl IntoRequest<UpdateCertificateRequest> ) -> Result<Response<Operation>, Status>

Updates a Certificate.

pub async fn delete_certificate( &mut self, request: impl IntoRequest<DeleteCertificateRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Certificate.

pub async fn list_certificate_maps( &mut self, request: impl IntoRequest<ListCertificateMapsRequest> ) -> Result<Response<ListCertificateMapsResponse>, Status>

Lists CertificateMaps in a given project and location.

pub async fn get_certificate_map( &mut self, request: impl IntoRequest<GetCertificateMapRequest> ) -> Result<Response<CertificateMap>, Status>

Gets details of a single CertificateMap.

pub async fn create_certificate_map( &mut self, request: impl IntoRequest<CreateCertificateMapRequest> ) -> Result<Response<Operation>, Status>

Creates a new CertificateMap in a given project and location.

pub async fn update_certificate_map( &mut self, request: impl IntoRequest<UpdateCertificateMapRequest> ) -> Result<Response<Operation>, Status>

Updates a CertificateMap.

pub async fn delete_certificate_map( &mut self, request: impl IntoRequest<DeleteCertificateMapRequest> ) -> Result<Response<Operation>, Status>

Deletes a single CertificateMap. A Certificate Map can’t be deleted if it contains Certificate Map Entries. Remove all the entries from the map before calling this method.

pub async fn list_certificate_map_entries( &mut self, request: impl IntoRequest<ListCertificateMapEntriesRequest> ) -> Result<Response<ListCertificateMapEntriesResponse>, Status>

Lists CertificateMapEntries in a given project and location.

pub async fn get_certificate_map_entry( &mut self, request: impl IntoRequest<GetCertificateMapEntryRequest> ) -> Result<Response<CertificateMapEntry>, Status>

Gets details of a single CertificateMapEntry.

pub async fn create_certificate_map_entry( &mut self, request: impl IntoRequest<CreateCertificateMapEntryRequest> ) -> Result<Response<Operation>, Status>

Creates a new CertificateMapEntry in a given project and location.

pub async fn update_certificate_map_entry( &mut self, request: impl IntoRequest<UpdateCertificateMapEntryRequest> ) -> Result<Response<Operation>, Status>

Updates a CertificateMapEntry.

pub async fn delete_certificate_map_entry( &mut self, request: impl IntoRequest<DeleteCertificateMapEntryRequest> ) -> Result<Response<Operation>, Status>

Deletes a single CertificateMapEntry.

pub async fn list_dns_authorizations( &mut self, request: impl IntoRequest<ListDnsAuthorizationsRequest> ) -> Result<Response<ListDnsAuthorizationsResponse>, Status>

Lists DnsAuthorizations in a given project and location.

pub async fn get_dns_authorization( &mut self, request: impl IntoRequest<GetDnsAuthorizationRequest> ) -> Result<Response<DnsAuthorization>, Status>

Gets details of a single DnsAuthorization.

pub async fn create_dns_authorization( &mut self, request: impl IntoRequest<CreateDnsAuthorizationRequest> ) -> Result<Response<Operation>, Status>

Creates a new DnsAuthorization in a given project and location.

pub async fn update_dns_authorization( &mut self, request: impl IntoRequest<UpdateDnsAuthorizationRequest> ) -> Result<Response<Operation>, Status>

Updates a DnsAuthorization.

pub async fn delete_dns_authorization( &mut self, request: impl IntoRequest<DeleteDnsAuthorizationRequest> ) -> Result<Response<Operation>, Status>

Deletes a single DnsAuthorization.

pub async fn list_certificate_issuance_configs( &mut self, request: impl IntoRequest<ListCertificateIssuanceConfigsRequest> ) -> Result<Response<ListCertificateIssuanceConfigsResponse>, Status>

Lists CertificateIssuanceConfigs in a given project and location.

pub async fn get_certificate_issuance_config( &mut self, request: impl IntoRequest<GetCertificateIssuanceConfigRequest> ) -> Result<Response<CertificateIssuanceConfig>, Status>

Gets details of a single CertificateIssuanceConfig.

pub async fn create_certificate_issuance_config( &mut self, request: impl IntoRequest<CreateCertificateIssuanceConfigRequest> ) -> Result<Response<Operation>, Status>

Creates a new CertificateIssuanceConfig in a given project and location.

pub async fn delete_certificate_issuance_config( &mut self, request: impl IntoRequest<DeleteCertificateIssuanceConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes a single CertificateIssuanceConfig.