Struct google_api_proto::google::cloud::channel::v1::Entitlement

pub struct Entitlement {

    pub name: String,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub offer: String,
    pub commitment_settings: Option<CommitmentSettings>,
    pub provisioning_state: i32,
    pub provisioned_service: Option<ProvisionedService>,
    pub suspension_reasons: Vec<i32>,
    pub purchase_order_id: String,
    pub trial_settings: Option<TrialSettings>,
    pub association_info: Option<AssociationInfo>,
    pub parameters: Vec<Parameter>,
    pub billing_account: String,

}

An entitlement is a representation of a customer’s ability to use a service.

Fields
name: String

Output only. Resource name of an entitlement in the form: accounts/{account_id}/customers/{customer_id}/entitlements/{entitlement_id}.

create_time: Option<Timestamp>

Output only. The time at which the entitlement is created.

update_time: Option<Timestamp>

Output only. The time at which the entitlement is updated.

offer: String

Required. The offer resource name for which the entitlement is to be created. Takes the form: accounts/{account_id}/offers/{offer_id}.

commitment_settings: Option<CommitmentSettings>

Commitment settings for a commitment-based Offer. Required for commitment based offers.

provisioning_state: i32

Output only. Current provisioning state of the entitlement.

provisioned_service: Option<ProvisionedService>

Output only. Service provisioning details for the entitlement.

suspension_reasons: Vec<i32>

Output only. Enumerable of all current suspension reasons for an entitlement.

purchase_order_id: String

Optional. This purchase order (PO) information is for resellers to use for their company tracking usage. If a purchaseOrderId value is given, it appears in the API responses and shows up in the invoice. The property accepts up to 80 plain text characters. This is only supported for Google Workspace entitlements.

trial_settings: Option<TrialSettings>

Output only. Settings for trial offers.

association_info: Option<AssociationInfo>

Association information to other entitlements.

parameters: Vec<Parameter>

Extended entitlement parameters. When creating an entitlement, valid parameter names and values are defined in the [Offer.parameter_definitions][google.cloud.channel.v1.Offer.parameter_definitions].

For Google Workspace, the following Parameters may be accepted as input:

max_units: The maximum assignable units for a flexible offer

OR

num_units: The total commitment for commitment-based offers

The response may additionally include the following output-only Parameters:

assigned_units: The number of licenses assigned to users.

For Google Cloud billing subaccounts, the following Parameter may be accepted as input:

display_name: The display name of the billing subaccount.

billing_account: String

Optional. The billing account resource name that is used to pay for this entitlement.

Implementations

impl Entitlement

pub fn provisioning_state(&self) -> ProvisioningState

Returns the enum value of provisioning_state, or the default if the field is set to an invalid enum value.

pub fn set_provisioning_state(&mut self, value: ProvisioningState)

Sets provisioning_state to the provided enum value.

pub fn suspension_reasons( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SuspensionReason>>

Returns an iterator which yields the valid enum values contained in suspension_reasons.

pub fn push_suspension_reasons(&mut self, value: SuspensionReason)

Appends the provided enum value to suspension_reasons.