Module google_api_proto::google::cloud::cloudcontrolspartner::v1

Modules

access_reason

Nested message and enum types in AccessReason.

cloud_controls_partner_core_client

Generated client implementations.

cloud_controls_partner_monitoring_client

Generated client implementations.

customer_onboarding_step

Nested message and enum types in CustomerOnboardingStep.

ekm_connection

Nested message and enum types in EkmConnection.

ekm_metadata

Nested message and enum types in EkmMetadata.

partner_permissions

Nested message and enum types in PartnerPermissions.

violation

Nested message and enum types in Violation.

workload

Nested message and enum types in Workload.

workload_onboarding_step

Nested message and enum types in WorkloadOnboardingStep.

Structs

AccessApprovalRequest

Details about the Access request.

AccessReason

Reason for the access.

Customer

Contains metadata around a Cloud Controls Partner Customer

CustomerOnboardingState

Container for customer onboarding steps

CustomerOnboardingStep

Container for customer onboarding information

EkmConnection

Details about the EKM connection

EkmConnections

The EKM connections associated with a workload

EkmMetadata

Holds information needed by Mudbray to use partner EKMs for workloads.

GetCustomerRequest

Message for getting a customer

GetEkmConnectionsRequest

Request for getting the EKM connections associated with a workload

GetPartnerPermissionsRequest

Request for getting the partner permissions granted for a workload

GetPartnerRequest

Message for getting a Partner

GetViolationRequest

Message for getting a Violation

GetWorkloadRequest

Message for getting a customer workload.

ListAccessApprovalRequestsRequest

Request for getting the access requests associated with a workload.

ListAccessApprovalRequestsResponse

Response message for list access requests.

ListCustomersRequest

Request to list customers

ListCustomersResponse

Response message for list customer Customers requests

ListViolationsRequest

Message for requesting list of Violations

ListViolationsResponse

Response message for list customer violation requests

ListWorkloadsRequest

Request to list customer workloads.

ListWorkloadsResponse

Response message for list customer workloads requests.

OperationMetadata

Represents the metadata of the long-running operation.

Partner

Message describing Partner resource

PartnerPermissions

The permissions granted to the partner for a workload

Sku

Represents the SKU a partner owns inside Google Cloud to sell to customers.

Violation

Details of resource Violation

Workload

Contains metadata around the Workload resource in the Assured Workloads API.

WorkloadOnboardingState

Container for workload onboarding steps.

WorkloadOnboardingStep

Container for workload onboarding information.

Enums

CompletionState

Enum for possible completion states.