Module google_api_proto::google::cloud::clouddms::v1

Modules

alloy_db_settings

Nested message and enum types in AlloyDbSettings.

apply_conversion_workspace_request

Nested message and enum types in ApplyConversionWorkspaceRequest.

apply_hash

Nested message and enum types in ApplyHash.

background_job_log_entry

Nested message and enum types in BackgroundJobLogEntry.

cloud_sql_settings

Nested message and enum types in CloudSqlSettings.

conditional_column_set_value

Nested message and enum types in ConditionalColumnSetValue.

connection_profile

Nested message and enum types in ConnectionProfile.

data_migration_service_client

Generated client implementations.

database_entity

Nested message and enum types in DatabaseEntity.

describe_database_entities_request

Nested message and enum types in DescribeDatabaseEntitiesRequest.

entity_issue

Nested message and enum types in EntityIssue.

forward_ssh_tunnel_connectivity

Nested message and enum types in ForwardSshTunnelConnectivity.

generate_ssh_script_request

Nested message and enum types in GenerateSshScriptRequest.

import_mapping_rules_request

Nested message and enum types in ImportMappingRulesRequest.

mapping_rule

Nested message and enum types in MappingRule.

migration_job

Nested message and enum types in MigrationJob.

migration_job_verification_error

Nested message and enum types in MigrationJobVerificationError.

multi_column_datatype_change

Nested message and enum types in MultiColumnDatatypeChange.

oracle_connection_profile

Nested message and enum types in OracleConnectionProfile.

postgre_sql_connection_profile

Nested message and enum types in PostgreSqlConnectionProfile.

private_connection

Nested message and enum types in PrivateConnection.

seed_conversion_workspace_request

Nested message and enum types in SeedConversionWorkspaceRequest.

sql_acl_entry

Nested message and enum types in SqlAclEntry.

ssl_config

Nested message and enum types in SslConfig.

value_transformation

Nested message and enum types in ValueTransformation.

Structs

AlloyDbConnectionProfile

Specifies required connection parameters, and the parameters required to create an AlloyDB destination cluster.

AlloyDbSettings

Settings for creating an AlloyDB cluster.

ApplyConversionWorkspaceRequest

Request message for ‘ApplyConversionWorkspace’ request.

ApplyHash

Apply a hash function on the value.

AssignSpecificValue

Set to a specific value (value is converted to fit the target data type)

BackgroundJobLogEntry

Execution log of a background job.

CloudSqlConnectionProfile

Specifies required connection parameters, and, optionally, the parameters required to create a Cloud SQL destination database instance.

CloudSqlSettings

Settings for creating a Cloud SQL database instance.

ColumnEntity

Column is not used as an independent entity, it is retrieved as part of a Table entity.

CommitConversionWorkspaceRequest

Request message for ‘CommitConversionWorkspace’ request.

ConditionalColumnSetValue

Options to configure rule type ConditionalColumnSetValue. The rule is used to transform the data which is being replicated/migrated.

ConnectionProfile

A connection profile definition.

ConstraintEntity

Constraint is not used as an independent entity, it is retrieved as part of another entity such as Table or View.

ConversionWorkspace

The main conversion workspace resource entity.

ConversionWorkspaceInfo

A conversion workspace’s version.

ConvertConversionWorkspaceRequest

Request message for ‘ConvertConversionWorkspace’ request.

ConvertRowIdToColumn

Options to configure rule type ConvertROWIDToColumn. The rule is used to add column rowid to destination tables based on an Oracle rowid function/property.

CreateConnectionProfileRequest

Request message for ‘CreateConnectionProfile’ request.

CreateConversionWorkspaceRequest

Request message to create a new Conversion Workspace in the specified project and region.

CreateMappingRuleRequest

Request message for ‘CreateMappingRule’ command.

CreateMigrationJobRequest

Request message to create a new Database Migration Service migration job in the specified project and region.

CreatePrivateConnectionRequest

Request message to create a new private connection in the specified project and region.

DatabaseEngineInfo

The type and version of a source or destination database.

DatabaseEntity

The base entity type for all the database related entities. The message contains the entity name, the name of its parent, the entity type, and the specific details per entity type.

DatabaseInstanceEntity

DatabaseInstance acts as a parent entity to other database entities.

DatabaseType

A message defining the database engine and provider.

DeleteConnectionProfileRequest

Request message for ‘DeleteConnectionProfile’ request.

DeleteConversionWorkspaceRequest

Request message for ‘DeleteConversionWorkspace’ request.

DeleteMappingRuleRequest

Request message for ‘DeleteMappingRule’ request.

DeleteMigrationJobRequest

Request message for ‘DeleteMigrationJob’ request.

DeletePrivateConnectionRequest

Request message to delete a private connection.

DescribeConversionWorkspaceRevisionsRequest

Request message for ‘DescribeConversionWorkspaceRevisions’ request.

DescribeConversionWorkspaceRevisionsResponse

Response message for ‘DescribeConversionWorkspaceRevisions’ request.

DescribeDatabaseEntitiesRequest

Request message for ‘DescribeDatabaseEntities’ request.

DescribeDatabaseEntitiesResponse

Response message for ‘DescribeDatabaseEntities’ request.

DoubleComparisonFilter

Filter based on relation between source value and compare value of type double in ConditionalColumnSetValue

EntityDdl

A single DDL statement for a specific entity

EntityIssue

Issue related to the entity.

EntityMapping

Details of the mappings of a database entity.

EntityMappingLogEntry

A single record of a rule which was used for a mapping.

EntityMove

Options to configure rule type EntityMove. The rule is used to move an entity to a new schema.

FetchStaticIpsRequest

Request message for ‘FetchStaticIps’ request.

FetchStaticIpsResponse

Response message for a ‘FetchStaticIps’ request.

FilterTableColumns

Options to configure rule type FilterTableColumns. The rule is used to filter the list of columns to include or exclude from a table.

ForwardSshTunnelConnectivity

Forward SSH Tunnel connectivity.

FunctionEntity

Function’s parent is a schema.

GenerateSshScriptRequest

Request message for ‘GenerateSshScript’ request.

GenerateTcpProxyScriptRequest

Request message for ‘GenerateTcpProxyScript’ request.

GetConnectionProfileRequest

Request message for ‘GetConnectionProfile’ request.

GetConversionWorkspaceRequest

Request message for ‘GetConversionWorkspace’ request.

GetMappingRuleRequest

Request message for ‘GetMappingRule’ request.

GetMigrationJobRequest

Request message for ‘GetMigrationJob’ request.

GetPrivateConnectionRequest

Request message to get a private connection resource.

ImportMappingRulesRequest

Request message for ‘ImportMappingRules’ request.

IndexEntity

Index is not used as an independent entity, it is retrieved as part of a Table entity.

IntComparisonFilter

Filter based on relation between source value and compare value of type integer in ConditionalColumnSetValue

ListConnectionProfilesRequest

Request message for ‘ListConnectionProfiles’ request.

ListConnectionProfilesResponse

Response message for ‘ListConnectionProfiles’ request.

ListConversionWorkspacesRequest

Retrieve a list of all conversion workspaces in a given project and location.

ListConversionWorkspacesResponse

Response message for ‘ListConversionWorkspaces’ request.

ListMappingRulesRequest

Retrieve a list of all mapping rules in a given conversion workspace.

ListMappingRulesResponse

Response message for ‘ListMappingRulesRequest’ request.

ListMigrationJobsRequest

Retrieves a list of all migration jobs in a given project and location.

ListMigrationJobsResponse

Response message for ‘ListMigrationJobs’ request.

ListPrivateConnectionsRequest

Request message to retrieve a list of private connections in a given project and location.

ListPrivateConnectionsResponse

Response message for ‘ListPrivateConnections’ request.

MappingRule

Definition of a transformation that is to be applied to a group of entities in the source schema. Several such transformations can be applied to an entity sequentially to define the corresponding entity in the target schema.

MappingRuleFilter

A filter defining the entities that a mapping rule should be applied to. When more than one field is specified, the rule is applied only to entities which match all the fields.

MaterializedViewEntity

MaterializedView’s parent is a schema.

MigrationJob

Represents a Database Migration Service migration job object.

MigrationJobVerificationError

Error message of a verification Migration job.

MultiColumnDatatypeChange

Options to configure rule type MultiColumnDatatypeChange. The rule is used to change the data type and associated properties of multiple columns at once.

MultiEntityRename

Options to configure rule type MultiEntityRename. The rule is used to rename multiple entities.

MySqlConnectionProfile

Specifies connection parameters required specifically for MySQL databases.

OperationMetadata

Represents the metadata of the long-running operation.

OracleConnectionProfile

Specifies connection parameters required specifically for Oracle databases.

PackageEntity

Package’s parent is a schema.

PostgreSqlConnectionProfile

Specifies connection parameters required specifically for PostgreSQL databases.

PrivateConnection

The PrivateConnection resource is used to establish private connectivity with the customer’s network.

PrivateConnectivity

Private Connectivity.

PrivateServiceConnectConnectivity

Private Service Connect connectivity

PromoteMigrationJobRequest

Request message for ‘PromoteMigrationJob’ request.

RestartMigrationJobRequest

Request message for ‘RestartMigrationJob’ request.

ResumeMigrationJobRequest

Request message for ‘ResumeMigrationJob’ request.

ReverseSshConnectivity

The details needed to configure a reverse SSH tunnel between the source and destination databases. These details will be used when calling the generateSshScript method (see https://cloud.google.com/database-migration/docs/reference/rest/v1/projects.locations.migrationJobs/generateSshScript) to produce the script that will help set up the reverse SSH tunnel, and to set up the VPC peering between the Cloud SQL private network and the VPC.

RollbackConversionWorkspaceRequest

Request message for ‘RollbackConversionWorkspace’ request.

RoundToScale

This allows the data to change scale, for example if the source is 2 digits after the decimal point, specify round to scale value = 2. If for example the value needs to be converted to an integer, use round to scale value = 0.

SchemaEntity

Schema typically has no parent entity, but can have a parent entity DatabaseInstance (for database engines which support it). For some database engines, the terms schema and user can be used interchangeably when they refer to a namespace or a collection of other database entities. Can store additional information which is schema specific.

SearchBackgroundJobsRequest

Request message for ‘SearchBackgroundJobs’ request.

SearchBackgroundJobsResponse

Response message for ‘SearchBackgroundJobs’ request.

SeedConversionWorkspaceRequest

Request message for ‘SeedConversionWorkspace’ request.

SequenceEntity

Sequence’s parent is a schema.

SetTablePrimaryKey

Options to configure rule type SetTablePrimaryKey. The rule is used to specify the columns and name to configure/alter the primary key of a table.

SingleColumnChange

Options to configure rule type SingleColumnChange. The rule is used to change the properties of a column.

SingleEntityRename

Options to configure rule type SingleEntityRename. The rule is used to rename an entity.

SinglePackageChange

Options to configure rule type SinglePackageChange. The rule is used to alter the sql code for a package entities.

SourceNumericFilter

Filter for fixed point number data types such as NUMERIC/NUMBER

SourceSqlChange

Options to configure rule type SourceSqlChange. The rule is used to alter the sql code for database entities.

SourceTextFilter

Filter for text-based data types like varchar.

SqlAclEntry

An entry for an Access Control list.

SqlIpConfig

IP Management configuration.

SshScript

Response message for ‘GenerateSshScript’ request.

SslConfig

SSL configuration information.

StartMigrationJobRequest

Request message for ‘StartMigrationJob’ request.

StaticIpConnectivity

The source database will allow incoming connections from the public IP of the destination database. You can retrieve the public IP of the Cloud SQL instance from the Cloud SQL console or using Cloud SQL APIs. No additional configuration is required.

StaticServiceIpConnectivity

Static IP address connectivity configured on service project.

StopMigrationJobRequest

Request message for ‘StopMigrationJob’ request.

StoredProcedureEntity

Stored procedure’s parent is a schema.

SynonymEntity

Synonym’s parent is a schema.

TableEntity

Table’s parent is a schema.

TcpProxyScript

Response message for ‘GenerateTcpProxyScript’ request.

TriggerEntity

Trigger is not used as an independent entity, it is retrieved as part of a Table entity.

UdtEntity

UDT’s parent is a schema.

UpdateConnectionProfileRequest

Request message for ‘UpdateConnectionProfile’ request.

UpdateConversionWorkspaceRequest

Request message for ‘UpdateConversionWorkspace’ request.

UpdateMigrationJobRequest

Request message for ‘UpdateMigrationJob’ request.

ValueListFilter

A list of values to filter by in ConditionalColumnSetValue

ValueTransformation

Description of data transformation during migration as part of the ConditionalColumnSetValue.

VerifyMigrationJobRequest

Request message for ‘VerifyMigrationJob’ request.

ViewEntity

View’s parent is a schema.

VmCreationConfig

VM creation configuration message

VmSelectionConfig

VM selection configuration message

VpcPeeringConfig

The VPC peering configuration is used to create VPC peering with the consumer’s VPC.

VpcPeeringConnectivity

The details of the VPC where the source database is located in Google Cloud. We will use this information to set up the VPC peering connection between Cloud SQL and this VPC.

Enums

BackgroundJobType

The types of jobs that can be executed in the background.

DatabaseEngine

The database engine types.

DatabaseEntityType

The type of database entities supported,

DatabaseEntityView

AIP-157 Partial Response view for Database Entity.

DatabaseProvider

The database providers.

EntityNameTransformation

Entity Name Transformation Types

ImportRulesFileFormat

The format for the import rules file.

NetworkArchitecture

NumericFilterOption

Specifies the columns on which numeric filter needs to be applied.

ValueComparison

Enum used by IntComparisonFilter and DoubleComparisonFilter to indicate the relation between source value and compare value.

ValuePresentInList

Enum used by ValueListFilter to indicate whether the source value is in the supplied list