Enum google_api_proto::google::cloud::clouddms::v1::migration_job_verification_error::ErrorCode

#[repr(i32)]
pub enum ErrorCode {

    Unspecified = 0,
    ConnectionFailure = 1,
    AuthenticationFailure = 2,
    InvalidConnectionProfileConfig = 3,
    VersionIncompatibility = 4,
    ConnectionProfileTypesIncompatibility = 5,
    NoPglogicalInstalled = 7,
    PglogicalNodeAlreadyExists = 8,
    InvalidWalLevel = 9,
    InvalidSharedPreloadLibrary = 10,
    InsufficientMaxReplicationSlots = 11,
    InsufficientMaxWalSenders = 12,
    InsufficientMaxWorkerProcesses = 13,
    UnsupportedExtensions = 14,
    UnsupportedMigrationType = 15,
    InvalidRdsLogicalReplication = 16,
    UnsupportedGtidMode = 17,
    UnsupportedTableDefinition = 18,
    UnsupportedDefiner = 19,
    CantRestartRunningMigration = 21,
    SourceAlreadySetup = 23,
    TablesWithLimitedSupport = 24,
    UnsupportedDatabaseLocale = 25,
    UnsupportedDatabaseFdwConfig = 26,
    ErrorRdbms = 27,
    SourceSizeExceedsThreshold = 28,
    ExistingConflictingDatabases = 29,
    ParallelImportInsufficientPrivilege = 30,

}

A general error code describing the type of error that occurred.

Variants

Unspecified = 0

An unknown error occurred

ConnectionFailure = 1

We failed to connect to one of the connection profile.

AuthenticationFailure = 2

We failed to authenticate to one of the connection profile.

InvalidConnectionProfileConfig = 3

One of the involved connection profiles has an invalid configuration.

VersionIncompatibility = 4

The versions of the source and the destination are incompatible.

ConnectionProfileTypesIncompatibility = 5

The types of the source and the destination are incompatible.

NoPglogicalInstalled = 7

No pglogical extension installed on databases, applicable for postgres.

PglogicalNodeAlreadyExists = 8

pglogical node already exists on databases, applicable for postgres.

InvalidWalLevel = 9

The value of parameter wal_level is not set to logical.

InvalidSharedPreloadLibrary = 10

The value of parameter shared_preload_libraries does not include pglogical.

InsufficientMaxReplicationSlots = 11

The value of parameter max_replication_slots is not sufficient.

InsufficientMaxWalSenders = 12

The value of parameter max_wal_senders is not sufficient.

InsufficientMaxWorkerProcesses = 13

The value of parameter max_worker_processes is not sufficient.

UnsupportedExtensions = 14

Extensions installed are either not supported or having unsupported versions.

UnsupportedMigrationType = 15

Unsupported migration type.

InvalidRdsLogicalReplication = 16

Invalid RDS logical replication.

UnsupportedGtidMode = 17

The gtid_mode is not supported, applicable for MySQL.

UnsupportedTableDefinition = 18

The table definition is not support due to missing primary key or replica identity.

UnsupportedDefiner = 19

The definer is not supported.

CantRestartRunningMigration = 21

Migration is already running at the time of restart request.

SourceAlreadySetup = 23

The source already has a replication setup.

TablesWithLimitedSupport = 24

The source has tables with limited support. E.g. PostgreSQL tables without primary keys.

UnsupportedDatabaseLocale = 25

The source uses an unsupported locale.

UnsupportedDatabaseFdwConfig = 26

The source uses an unsupported Foreign Data Wrapper configuration.

ErrorRdbms = 27

There was an underlying RDBMS error.

SourceSizeExceedsThreshold = 28

The source DB size in Bytes exceeds a certain threshold. The migration might require an increase of quota, or might not be supported.

ExistingConflictingDatabases = 29

The destination DB contains existing databases that are conflicting with those in the source DB.

ParallelImportInsufficientPrivilege = 30

Insufficient privilege to enable the parallelism configuration.

Implementations

impl ErrorCode

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ErrorCode.

pub fn from_i32(value: i32) -> Option<ErrorCode>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ErrorCode, or None if value is not a valid variant.

impl ErrorCode

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.