Struct google_api_proto::google::cloud::clouddms::v1::CloudSqlSettings

pub struct CloudSqlSettings {

    pub database_version: i32,
    pub user_labels: BTreeMap<String, String>,
    pub tier: String,
    pub storage_auto_resize_limit: Option<i64>,
    pub activation_policy: i32,
    pub ip_config: Option<SqlIpConfig>,
    pub auto_storage_increase: Option<bool>,
    pub database_flags: BTreeMap<String, String>,
    pub data_disk_type: i32,
    pub data_disk_size_gb: Option<i64>,
    pub zone: String,
    pub secondary_zone: String,
    pub source_id: String,
    pub root_password: String,
    pub root_password_set: bool,
    pub collation: String,
    pub cmek_key_name: String,
    pub availability_type: i32,
    pub edition: i32,

}

Settings for creating a Cloud SQL database instance.

Fields
database_version: i32

The database engine type and version.

user_labels: BTreeMap<String, String>

The resource labels for a Cloud SQL instance to use to annotate any related underlying resources such as Compute Engine VMs. An object containing a list of “key”: “value” pairs.

Example: { "name": "wrench", "mass": "18kg", "count": "3" }.

tier: String

The tier (or machine type) for this instance, for example: db-n1-standard-1 (MySQL instances) or db-custom-1-3840 (PostgreSQL instances). For more information, see Cloud SQL Instance Settings.

storage_auto_resize_limit: Option<i64>

The maximum size to which storage capacity can be automatically increased. The default value is 0, which specifies that there is no limit.

activation_policy: i32

The activation policy specifies when the instance is activated; it is applicable only when the instance state is ‘RUNNABLE’. Valid values:

‘ALWAYS’: The instance is on, and remains so even in the absence of connection requests.

NEVER: The instance is off; it is not activated, even if a connection request arrives.

ip_config: Option<SqlIpConfig>

The settings for IP Management. This allows to enable or disable the instance IP and manage which external networks can connect to the instance. The IPv4 address cannot be disabled.

auto_storage_increase: Option<bool>

[default: ON] If you enable this setting, Cloud SQL checks your available storage every 30 seconds. If the available storage falls below a threshold size, Cloud SQL automatically adds additional storage capacity. If the available storage repeatedly falls below the threshold size, Cloud SQL continues to add storage until it reaches the maximum of 30 TB.

database_flags: BTreeMap<String, String>

The database flags passed to the Cloud SQL instance at startup. An object containing a list of “key”: value pairs. Example: { “name”: “wrench”, “mass”: “1.3kg”, “count”: “3” }.

data_disk_type: i32

The type of storage: PD_SSD (default) or PD_HDD.

data_disk_size_gb: Option<i64>

The storage capacity available to the database, in GB. The minimum (and default) size is 10GB.

zone: String

The Google Cloud Platform zone where your Cloud SQL database instance is located.

secondary_zone: String

Optional. The Google Cloud Platform zone where the failover Cloud SQL database instance is located. Used when the Cloud SQL database availability type is REGIONAL (i.e. multiple zones / highly available).

source_id: String

The Database Migration Service source connection profile ID, in the format: projects/my_project_name/locations/us-central1/connectionProfiles/connection_profile_ID

root_password: String

Input only. Initial root password.

root_password_set: bool

Output only. Indicates If this connection profile root password is stored.

collation: String

The Cloud SQL default instance level collation.

cmek_key_name: String

The KMS key name used for the csql instance.

availability_type: i32

Optional. Availability type. Potential values:

ZONAL: The instance serves data from only one zone. Outages in that zone affect data availability.

REGIONAL: The instance can serve data from more than one zone in a region (it is highly available).

edition: i32

Optional. The edition of the given Cloud SQL instance.

Implementations

impl CloudSqlSettings

pub fn database_version(&self) -> SqlDatabaseVersion

Returns the enum value of database_version, or the default if the field is set to an invalid enum value.

pub fn set_database_version(&mut self, value: SqlDatabaseVersion)

Sets database_version to the provided enum value.

pub fn activation_policy(&self) -> SqlActivationPolicy

Returns the enum value of activation_policy, or the default if the field is set to an invalid enum value.

pub fn set_activation_policy(&mut self, value: SqlActivationPolicy)

Sets activation_policy to the provided enum value.

pub fn data_disk_type(&self) -> SqlDataDiskType

Returns the enum value of data_disk_type, or the default if the field is set to an invalid enum value.

pub fn set_data_disk_type(&mut self, value: SqlDataDiskType)

Sets data_disk_type to the provided enum value.

pub fn availability_type(&self) -> SqlAvailabilityType

Returns the enum value of availability_type, or the default if the field is set to an invalid enum value.

pub fn set_availability_type(&mut self, value: SqlAvailabilityType)

Sets availability_type to the provided enum value.

pub fn edition(&self) -> Edition

Returns the enum value of edition, or the default if the field is set to an invalid enum value.

pub fn set_edition(&mut self, value: Edition)

Sets edition to the provided enum value.