Module google_api_proto::google::cloud::commerce::consumer::procurement::v1

Modules

consumer_procurement_service_client

Generated client implementations.

parameter

Nested message and enum types in Parameter.

Structs

GetOrderRequest

Request message for [ConsumerProcurementService.GetOrder][google.cloud.commerce.consumer.procurement.v1.ConsumerProcurementService.GetOrder]

LineItem

A single item within an order.

LineItemChange

A change made on a line item.

LineItemInfo

Line item information.

ListOrdersRequest

Request message for [ConsumerProcurementService.ListOrders][google.cloud.commerce.consumer.procurement.v1.ConsumerProcurementService.ListOrders].

ListOrdersResponse

Response message for [ConsumerProcurementService.ListOrders][google.cloud.commerce.consumer.procurement.v1.ConsumerProcurementService.ListOrders].

Order

Represents a purchase made by a customer on Cloud Marketplace. Creating an order makes sure that both the Google backend systems as well as external service provider’s systems (if needed) allow use of purchased products and ensures the appropriate billing events occur.

Parameter

User-provided Parameters.

PlaceOrderMetadata

Message stored in the metadata field of the Operation returned by [ConsumerProcurementService.PlaceOrder][google.cloud.commerce.consumer.procurement.v1.ConsumerProcurementService.PlaceOrder].

PlaceOrderRequest

Request message for [ConsumerProcurementService.PlaceOrder][google.cloud.commerce.consumer.procurement.v1.ConsumerProcurementService.PlaceOrder].

Subscription

Subscription information.

Enums

LineItemChangeState

State of a change.

LineItemChangeStateReasonType

Predefined types for line item change state reason.

LineItemChangeType

Type of a line item change.