Enum google_api_proto::google::cloud::compute::v1::autoscaler_status_details::Type

#[repr(i32)]
pub enum Type {

    UndefinedType = 0,
    AllInstancesUnhealthy = 404_965_477,
    BackendServiceDoesNotExist = 191_417_626,
    CappedAtMaxNumReplicas = 518_617,
    CustomMetricDataPointsTooSparse = 328_964_659,
    CustomMetricInvalid = 204_430_550,
    MinEqualsMax = 2_821_361,
    MissingCustomMetricDataPoints = 94_885_086,
    MissingLoadBalancingDataPoints = 509_858_898,
    ModeOff = 164_169_907,
    ModeOnlyScaleOut = 3_840_994,
    ModeOnlyUp = 100_969_842,
    MoreThanOneBackendService = 151_922_141,
    NotEnoughQuotaAvailable = 403_101_631,
    RegionResourceStockout = 528_622_846,
    ScalingTargetDoesNotExist = 122_636_699,
    ScheduledInstancesGreaterThanAutoscalerMax = 29_275_586,
    ScheduledInstancesLessThanAutoscalerMin = 398_287_669,
    Unknown = 433_141_802,
    UnsupportedMaxRateLoadBalancingConfiguration = 330_845_009,
    ZoneResourceStockout = 210_200_502,

}

The type of error, warning, or notice returned. Current set of possible values: - ALL_INSTANCES_UNHEALTHY (WARNING): All instances in the instance group are unhealthy (not in RUNNING state). - BACKEND_SERVICE_DOES_NOT_EXIST (ERROR): There is no backend service attached to the instance group. - CAPPED_AT_MAX_NUM_REPLICAS (WARNING): Autoscaler recommends a size greater than maxNumReplicas. - CUSTOM_METRIC_DATA_POINTS_TOO_SPARSE (WARNING): The custom metric samples are not exported often enough to be a credible base for autoscaling. - CUSTOM_METRIC_INVALID (ERROR): The custom metric that was specified does not exist or does not have the necessary labels. - MIN_EQUALS_MAX (WARNING): The minNumReplicas is equal to maxNumReplicas. This means the autoscaler cannot add or remove instances from the instance group. - MISSING_CUSTOM_METRIC_DATA_POINTS (WARNING): The autoscaler did not receive any data from the custom metric configured for autoscaling. - MISSING_LOAD_BALANCING_DATA_POINTS (WARNING): The autoscaler is configured to scale based on a load balancing signal but the instance group has not received any requests from the load balancer. - MODE_OFF (WARNING): Autoscaling is turned off. The number of instances in the group won’t change automatically. The autoscaling configuration is preserved. - MODE_ONLY_UP (WARNING): Autoscaling is in the “Autoscale only out” mode. The autoscaler can add instances but not remove any. - MORE_THAN_ONE_BACKEND_SERVICE (ERROR): The instance group cannot be autoscaled because it has more than one backend service attached to it. - NOT_ENOUGH_QUOTA_AVAILABLE (ERROR): There is insufficient quota for the necessary resources, such as CPU or number of instances. - REGION_RESOURCE_STOCKOUT (ERROR): Shown only for regional autoscalers: there is a resource stockout in the chosen region. - SCALING_TARGET_DOES_NOT_EXIST (ERROR): The target to be scaled does not exist. - UNSUPPORTED_MAX_RATE_LOAD_BALANCING_CONFIGURATION (ERROR): Autoscaling does not work with an HTTP/S load balancer that has been configured for maxRate. - ZONE_RESOURCE_STOCKOUT (ERROR): For zonal autoscalers: there is a resource stockout in the chosen zone. For regional autoscalers: in at least one of the zones you’re using there is a resource stockout. New values might be added in the future. Some of the values might not be available in all API versions.

Variants

UndefinedType = 0

A value indicating that the enum field is not set.

AllInstancesUnhealthy = 404_965_477

All instances in the instance group are unhealthy (not in RUNNING state).

BackendServiceDoesNotExist = 191_417_626

There is no backend service attached to the instance group.

CappedAtMaxNumReplicas = 518_617

Autoscaler recommends a size greater than maxNumReplicas.

CustomMetricDataPointsTooSparse = 328_964_659

The custom metric samples are not exported often enough to be a credible base for autoscaling.

CustomMetricInvalid = 204_430_550

The custom metric that was specified does not exist or does not have the necessary labels.

MinEqualsMax = 2_821_361

The minNumReplicas is equal to maxNumReplicas. This means the autoscaler cannot add or remove instances from the instance group.

MissingCustomMetricDataPoints = 94_885_086

The autoscaler did not receive any data from the custom metric configured for autoscaling.

MissingLoadBalancingDataPoints = 509_858_898

The autoscaler is configured to scale based on a load balancing signal but the instance group has not received any requests from the load balancer.

ModeOff = 164_169_907

Autoscaling is turned off. The number of instances in the group won’t change automatically. The autoscaling configuration is preserved.

ModeOnlyScaleOut = 3_840_994

Autoscaling is in the “Autoscale only scale out” mode. Instances in the group will be only added.

ModeOnlyUp = 100_969_842

Autoscaling is in the “Autoscale only out” mode. Instances in the group will be only added.

MoreThanOneBackendService = 151_922_141

The instance group cannot be autoscaled because it has more than one backend service attached to it.

NotEnoughQuotaAvailable = 403_101_631

There is insufficient quota for the necessary resources, such as CPU or number of instances.

RegionResourceStockout = 528_622_846

Showed only for regional autoscalers: there is a resource stockout in the chosen region.

ScalingTargetDoesNotExist = 122_636_699

The target to be scaled does not exist.

ScheduledInstancesGreaterThanAutoscalerMax = 29_275_586

For some scaling schedules minRequiredReplicas is greater than maxNumReplicas. Autoscaler always recommends at most maxNumReplicas instances.

ScheduledInstancesLessThanAutoscalerMin = 398_287_669

For some scaling schedules minRequiredReplicas is less than minNumReplicas. Autoscaler always recommends at least minNumReplicas instances.

Unknown = 433_141_802

UnsupportedMaxRateLoadBalancingConfiguration = 330_845_009

Autoscaling does not work with an HTTP/S load balancer that has been configured for maxRate.

ZoneResourceStockout = 210_200_502

For zonal autoscalers: there is a resource stockout in the chosen zone. For regional autoscalers: in at least one of the zones you’re using there is a resource stockout.

Implementations

impl Type

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Type.

pub fn from_i32(value: i32) -> Option<Type>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Type, or None if value is not a valid variant.

impl Type

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.