Module google_api_proto::google::cloud::compute::v1::instance_group_manager_update_policy

Nested message and enum types in InstanceGroupManagerUpdatePolicy.

Enums

InstanceRedistributionType

The instance redistribution policy for regional managed instance groups. Valid values are: - PROACTIVE (default): The group attempts to maintain an even distribution of VM instances across zones in the region. - NONE: For non-autoscaled groups, proactive redistribution is disabled. Additional supported values which may be not listed in the enum directly due to technical reasons: NONE PROACTIVE

MinimalAction

Minimal action to be taken on an instance. Use this option to minimize disruption as much as possible or to apply a more disruptive action than is necessary. - To limit disruption as much as possible, set the minimal action to REFRESH. If your update requires a more disruptive action, Compute Engine performs the necessary action to execute the update. - To apply a more disruptive action than is strictly necessary, set the minimal action to RESTART or REPLACE. For example, Compute Engine does not need to restart a VM to change its metadata. But if your application reads instance metadata only when a VM is restarted, you can set the minimal action to RESTART in order to pick up metadata changes. Additional supported values which may be not listed in the enum directly due to technical reasons: NONE REFRESH REPLACE RESTART

MostDisruptiveAllowedAction

Most disruptive action that is allowed to be taken on an instance. You can specify either NONE to forbid any actions, REFRESH to avoid restarting the VM and to limit disruption as much as possible. RESTART to allow actions that can be applied without instance replacing or REPLACE to allow all possible actions. If the Updater determines that the minimal update action needed is more disruptive than most disruptive allowed action you specify it will not perform the update at all. Additional supported values which may be not listed in the enum directly due to technical reasons: NONE REFRESH REPLACE RESTART

ReplacementMethod

What action should be used to replace instances. See minimal_action.REPLACE

Type

The type of update process. You can specify either PROACTIVE so that the MIG automatically updates VMs to the latest configurations or OPPORTUNISTIC so that you can select the VMs that you want to update. Additional supported values which may be not listed in the enum directly due to technical reasons: PROACTIVE