Struct google_api_proto::google::cloud::compute::v1::instance_group_managers_client::InstanceGroupManagersClient

pub struct InstanceGroupManagersClient<T> { /* private fields */ }

The InstanceGroupManagers API.

Implementations

impl<T> InstanceGroupManagersClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> InstanceGroupManagersClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn abandon_instances( &mut self, request: impl IntoRequest<AbandonInstancesInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified instances to be removed from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn aggregated_list( &mut self, request: impl IntoRequest<AggregatedListInstanceGroupManagersRequest> ) -> Result<Response<InstanceGroupManagerAggregatedList>, Status>

Retrieves the list of managed instance groups and groups them by zone.

pub async fn apply_updates_to_instances( &mut self, request: impl IntoRequest<ApplyUpdatesToInstancesInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Applies changes to selected instances on the managed instance group. This method can be used to apply new overrides and/or new versions.

pub async fn create_instances( &mut self, request: impl IntoRequest<CreateInstancesInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Creates instances with per-instance configurations in this managed instance group. Instances are created using the current instance template. The create instances operation is marked DONE if the createInstances request is successful. The underlying actions take additional time. You must separately verify the status of the creating or actions with the listmanagedinstances method.

pub async fn delete( &mut self, request: impl IntoRequest<DeleteInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified managed instance group and all of the instances in that group. Note that the instance group must not belong to a backend service. Read Deleting an instance group for more information.

pub async fn delete_instances( &mut self, request: impl IntoRequest<DeleteInstancesInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified instances in the managed instance group for immediate deletion. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. This operation is marked as DONE when the action is scheduled even if the instances are still being deleted. You must separately verify the status of the deleting action with the listmanagedinstances method. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn delete_per_instance_configs( &mut self, request: impl IntoRequest<DeletePerInstanceConfigsInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Deletes selected per-instance configurations for the managed instance group.

pub async fn get( &mut self, request: impl IntoRequest<GetInstanceGroupManagerRequest> ) -> Result<Response<InstanceGroupManager>, Status>

Returns all of the details about the specified managed instance group.

pub async fn insert( &mut self, request: impl IntoRequest<InsertInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Creates a managed instance group using the information that you specify in the request. After the group is created, instances in the group are created using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method. A managed instance group can have up to 1000 VM instances per group. Please contact Cloud Support if you need an increase in this limit.

pub async fn list( &mut self, request: impl IntoRequest<ListInstanceGroupManagersRequest> ) -> Result<Response<InstanceGroupManagerList>, Status>

Retrieves a list of managed instance groups that are contained within the specified project and zone.

pub async fn list_errors( &mut self, request: impl IntoRequest<ListErrorsInstanceGroupManagersRequest> ) -> Result<Response<InstanceGroupManagersListErrorsResponse>, Status>

Lists all errors thrown by actions on instances for a given managed instance group. The filter and orderBy query parameters are not supported.

pub async fn list_managed_instances( &mut self, request: impl IntoRequest<ListManagedInstancesInstanceGroupManagersRequest> ) -> Result<Response<InstanceGroupManagersListManagedInstancesResponse>, Status>

Lists all of the instances in the managed instance group. Each instance in the list has a currentAction, which indicates the action that the managed instance group is performing on the instance. For example, if the group is still creating an instance, the currentAction is CREATING. If a previous action failed, the list displays the errors for that failed action. The orderBy query parameter is not supported. The pageToken query parameter is supported only in the alpha and beta API and only if the group’s listManagedInstancesResults field is set to PAGINATED.

pub async fn list_per_instance_configs( &mut self, request: impl IntoRequest<ListPerInstanceConfigsInstanceGroupManagersRequest> ) -> Result<Response<InstanceGroupManagersListPerInstanceConfigsResp>, Status>

Lists all of the per-instance configurations defined for the managed instance group. The orderBy query parameter is not supported.

pub async fn patch( &mut self, request: impl IntoRequest<PatchInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Updates a managed instance group using the information that you specify in the request. This operation is marked as DONE when the group is patched even if the instances in the group are still in the process of being patched. You must separately verify the status of the individual instances with the listManagedInstances method. This method supports PATCH semantics and uses the JSON merge patch format and processing rules. If you update your group to specify a new template or instance configuration, it’s possible that your intended specification for each VM in the group is different from the current state of that VM. To learn how to apply an updated configuration to the VMs in a MIG, see Updating instances in a MIG.

pub async fn patch_per_instance_configs( &mut self, request: impl IntoRequest<PatchPerInstanceConfigsInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Inserts or patches per-instance configurations for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.

pub async fn recreate_instances( &mut self, request: impl IntoRequest<RecreateInstancesInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified VM instances in the managed instance group to be immediately recreated. Each instance is recreated using the group’s current configuration. This operation is marked as DONE when the flag is set even if the instances have not yet been recreated. You must separately verify the status of each instance by checking its currentAction field; for more information, see Checking the status of managed instances. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn resize( &mut self, request: impl IntoRequest<ResizeInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Resizes the managed instance group. If you increase the size, the group creates new instances using the current instance template. If you decrease the size, the group deletes instances. The resize operation is marked DONE when the resize actions are scheduled even if the group has not yet added or deleted any instances. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method. When resizing down, the instance group arbitrarily chooses the order in which VMs are deleted. The group takes into account some VM attributes when making the selection including: + The status of the VM instance. + The health of the VM instance. + The instance template version the VM is based on. + For regional managed instance groups, the location of the VM instance. This list is subject to change. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

pub async fn set_instance_template( &mut self, request: impl IntoRequest<SetInstanceTemplateInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Specifies the instance template to use when creating new instances in this group. The templates for existing instances in the group do not change unless you run recreateInstances, run applyUpdatesToInstances, or set the group’s updatePolicy.type to PROACTIVE.

pub async fn set_target_pools( &mut self, request: impl IntoRequest<SetTargetPoolsInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Modifies the target pools to which all instances in this managed instance group are assigned. The target pools automatically apply to all of the instances in the managed instance group. This operation is marked DONE when you make the request even if the instances have not yet been added to their target pools. The change might take some time to apply to all of the instances in the group depending on the size of the group.

pub async fn update_per_instance_configs( &mut self, request: impl IntoRequest<UpdatePerInstanceConfigsInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Inserts or updates per-instance configurations for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.