Enum google_api_proto::google::cloud::compute::v1::quota::Metric

#[repr(i32)]
pub enum Metric {

    UndefinedMetric = 0,
    A2Cpus = 153_206_585,
    AffinityGroups = 108_303_563,
    Autoscalers = 471_248_988,
    BackendBuckets = 137_626_846,
    BackendServices = 269_623_753,
    C2dCpus = 508_182_517,
    C2Cpus = 317_601_211,
    C3Cpus = 346_230_362,
    Commitments = 456_141_790,
    CommittedA2Cpus = 59_330_902,
    CommittedC2dCpus = 282_390_904,
    CommittedC2Cpus = 223_725_528,
    CommittedC3Cpus = 252_354_679,
    CommittedCpus = 292_394_702,
    CommittedE2Cpus = 388_120_154,
    CommittedLicenses = 357_606_869,
    CommittedLocalSsdTotalGb = 308_393_480,
    CommittedM3Cpus = 585_985,
    CommittedMemoryOptimizedCpus = 489_057_886,
    CommittedN2aCpus = 40_064_304,
    CommittedN2dCpus = 125_951_757,
    CommittedN2Cpus = 322_589_603,
    CommittedNvidiaA10080gbGpus = 464_326_565,
    CommittedNvidiaA100Gpus = 375_799_445,
    CommittedNvidiaH100Gpus = 71_649_180,
    CommittedNvidiaK80Gpus = 3_857_188,
    CommittedNvidiaL4Gpus = 19_163_645,
    CommittedNvidiaP100Gpus = 107_528_100,
    CommittedNvidiaP4Gpus = 347_952_897,
    CommittedNvidiaT4Gpus = 139_871_237,
    CommittedNvidiaV100Gpus = 219_562,
    CommittedT2aCpus = 296_378_986,
    CommittedT2dCpus = 382_266_439,
    CommittedZ3Cpus = 263_844_686,
    Cpus = 2_075_595,
    CpusAllRegions = 470_911_149,
    DisksTotalGb = 353_520_543,
    E2Cpus = 481_995_837,
    ExternalManagedForwardingRules = 150_790_089,
    ExternalNetworkLbForwardingRules = 374_298_265,
    ExternalProtocolForwardingRules = 63_478_888,
    ExternalVpnGateways = 272_457_134,
    Firewalls = 374_485_843,
    ForwardingRules = 432_668_949,
    GlobalExternalManagedBackendServices = 164_566_753,
    GlobalExternalManagedForwardingRules = 327_611_949,
    GlobalExternalProxyLbBackendServices = 400_256_169,
    GlobalInternalAddresses = 42_738_332,
    GlobalInternalManagedBackendServices = 256_608_303,
    GlobalInternalTrafficDirectorBackendServices = 323_514_196,
    GpusAllRegions = 39_387_177,
    HdbTotalGb = 319_316_271,
    HdbTotalIops = 309_720_317,
    HdbTotalThroughput = 20_981_374,
    HealthChecks = 289_347_502,
    Images = 15_562_360,
    Instances = 131_337_822,
    InstanceGroups = 355_919_038,
    InstanceGroupManagers = 101_798_192,
    InstanceTemplates = 226_188_271,
    Interconnects = 415_204_741,
    InterconnectAttachmentsPerRegion = 159_968_086,
    InterconnectAttachmentsTotalMbps = 425_090_419,
    InterconnectTotalGbps = 285_341_866,
    InternalAddresses = 197_899_392,
    InternalTrafficDirectorForwardingRules = 266_433_668,
    InPlaceSnapshots = 151_359_133,
    InUseAddresses = 402_125_072,
    InUseBackupSchedules = 32_786_705,
    InUseSnapshotSchedules = 462_104_083,
    LocalSsdTotalGb = 330_878_021,
    M1Cpus = 37_203_366,
    M2Cpus = 65_832_517,
    M3Cpus = 94_461_668,
    MachineImages = 446_986_640,
    N2aCpus = 265_855_917,
    N2dCpus = 351_743_370,
    N2Cpus = 416_465_286,
    Networks = 485_481_477,
    NetworkAttachments = 149_028_575,
    NetworkEndpointGroups = 102_144_909,
    NetworkFirewallPolicies = 101_117_374,
    NetLbSecurityPoliciesPerRegion = 157_892_269,
    NetLbSecurityPolicyRulesPerRegion = 356_090_931,
    NetLbSecurityPolicyRuleAttributesPerRegion = 311_243_888,
    NodeGroups = 24_624_817,
    NodeTemplates = 474_896_668,
    NvidiaA10080gbGpus = 286_389_320,
    NvidiaA100Gpus = 504_872_978,
    NvidiaK80Gpus = 163_886_599,
    NvidiaL4Gpus = 491_923_130,
    NvidiaP100Gpus = 236_601_633,
    NvidiaP100VwsGpus = 213_970_574,
    NvidiaP4Gpus = 283_841_470,
    NvidiaP4VwsGpus = 528_296_619,
    NvidiaT4Gpus = 75_759_810,
    NvidiaT4VwsGpus = 319_813_039,
    NvidiaV100Gpus = 129_293_095,
    PacketMirrorings = 15_578_407,
    PdExtremeTotalProvisionedIops = 69_593_965,
    PreemptibleCpus = 251_184_841,
    PreemptibleLocalSsdGb = 260_819_336,
    PreemptibleNvidiaA10080gbGpus = 151_942_410,
    PreemptibleNvidiaA100Gpus = 68_832_784,
    PreemptibleNvidiaH100Gpus = 301_553_431,
    PreemptibleNvidiaK80Gpus = 374_960_201,
    PreemptibleNvidiaL4Gpus = 100_408_376,
    PreemptibleNvidiaP100Gpus = 337_432_351,
    PreemptibleNvidiaP100VwsGpus = 313_544_076,
    PreemptibleNvidiaP4Gpus = 429_197_628,
    PreemptibleNvidiaP4VwsGpus = 252_981_545,
    PreemptibleNvidiaT4Gpus = 221_115_968,
    PreemptibleNvidiaT4VwsGpus = 44_497_965,
    PreemptibleNvidiaV100Gpus = 230_123_813,
    PreemptibleTpuLiteDeviceV5 = 174_262_523,
    PreemptibleTpuLitePodsliceV5 = 112_281_796,
    PreemptibleTpuPodsliceV4 = 425_541_132,
    PscIlbConsumerForwardingRulesPerProducerNetwork = 231_164_291,
    PscInternalLbForwardingRules = 169_005_435,
    PublicAdvertisedPrefixes = 471_371_980,
    PublicDelegatedPrefixes = 532_465_974,
    RegionalAutoscalers = 29_363_772,
    RegionalExternalManagedBackendServices = 4_240_989,
    RegionalExternalNetworkLbBackendServices = 409_564_525,
    RegionalInstanceGroupManagers = 37_543_696,
    RegionalInternalLbBackendServices = 137_983_760,
    RegionalInternalManagedBackendServices = 96_282_539,
    RegionalInternalTrafficDirectorBackendServices = 483_162_968,
    Reservations = 32_644_647,
    ResourcePolicies = 83_955_297,
    Routers = 493_018_666,
    Routes = 275_680_074,
    SecurityPolicies = 189_518_703,
    SecurityPoliciesPerRegion = 249_041_734,
    SecurityPolicyAdvancedRulesPerRegion = 371_815_341,
    SecurityPolicyCevalRules = 470_815_689,
    SecurityPolicyRules = 203_549_225,
    SecurityPolicyRulesPerRegion = 126_510_156,
    ServiceAttachments = 471_521_510,
    Snapshots = 343_405_327,
    SsdTotalGb = 161_732_561,
    SslCertificates = 378_372_399,
    SslPolicies = 523_254_339,
    StaticAddresses = 93_624_049,
    StaticByoipAddresses = 275_809_649,
    StaticExternalIpv6AddressRanges = 472_346_774,
    Subnetworks = 421_330_469,
    T2aCpus = 522_170_599,
    T2dCpus = 71_187_140,
    TargetHttpsProxies = 219_522_506,
    TargetHttpProxies = 164_117_155,
    TargetInstances = 284_519_728,
    TargetPools = 348_261_257,
    TargetSslProxies = 159_216_235,
    TargetTcpProxies = 182_243_136,
    TargetVpnGateways = 75_029_928,
    TpuLiteDeviceV5 = 449_577_597,
    TpuLitePodsliceV5 = 12_708_294,
    TpuPodsliceV4 = 214_467_530,
    UrlMaps = 378_660_743,
    VariableIpv6PublicDelegatedPrefixes = 128_400_161,
    VpnGateways = 35_620_282,
    VpnTunnels = 104_327_296,
    XpnServiceProjects = 95_191_981,

}

[Output Only] Name of the quota metric.

Variants

UndefinedMetric = 0

A value indicating that the enum field is not set.

A2Cpus = 153_206_585

AffinityGroups = 108_303_563

Autoscalers = 471_248_988

BackendBuckets = 137_626_846

BackendServices = 269_623_753

C2dCpus = 508_182_517

C2Cpus = 317_601_211

C3Cpus = 346_230_362

Commitments = 456_141_790

CommittedA2Cpus = 59_330_902

CommittedC2dCpus = 282_390_904

CommittedC2Cpus = 223_725_528

CommittedC3Cpus = 252_354_679

CommittedCpus = 292_394_702

CommittedE2Cpus = 388_120_154

CommittedLicenses = 357_606_869

CommittedLocalSsdTotalGb = 308_393_480

CommittedM3Cpus = 585_985

CommittedMemoryOptimizedCpus = 489_057_886

CommittedN2aCpus = 40_064_304

CommittedN2dCpus = 125_951_757

CommittedN2Cpus = 322_589_603

CommittedNvidiaA10080gbGpus = 464_326_565

CommittedNvidiaA100Gpus = 375_799_445

CommittedNvidiaH100Gpus = 71_649_180

CommittedNvidiaK80Gpus = 3_857_188

CommittedNvidiaL4Gpus = 19_163_645

CommittedNvidiaP100Gpus = 107_528_100

CommittedNvidiaP4Gpus = 347_952_897

CommittedNvidiaT4Gpus = 139_871_237

CommittedNvidiaV100Gpus = 219_562

CommittedT2aCpus = 296_378_986

CommittedT2dCpus = 382_266_439

CommittedZ3Cpus = 263_844_686

Cpus = 2_075_595

Guest CPUs

CpusAllRegions = 470_911_149

DisksTotalGb = 353_520_543

E2Cpus = 481_995_837

ExternalManagedForwardingRules = 150_790_089

ExternalNetworkLbForwardingRules = 374_298_265

ExternalProtocolForwardingRules = 63_478_888

ExternalVpnGateways = 272_457_134

Firewalls = 374_485_843

ForwardingRules = 432_668_949

GlobalExternalManagedBackendServices = 164_566_753

GlobalExternalManagedForwardingRules = 327_611_949

GlobalExternalProxyLbBackendServices = 400_256_169

GlobalInternalAddresses = 42_738_332

GlobalInternalManagedBackendServices = 256_608_303

GlobalInternalTrafficDirectorBackendServices = 323_514_196

GpusAllRegions = 39_387_177

HdbTotalGb = 319_316_271

HdbTotalIops = 309_720_317

HdbTotalThroughput = 20_981_374

HealthChecks = 289_347_502

Images = 15_562_360

Instances = 131_337_822

InstanceGroups = 355_919_038

InstanceGroupManagers = 101_798_192

InstanceTemplates = 226_188_271

Interconnects = 415_204_741

InterconnectAttachmentsPerRegion = 159_968_086

InterconnectAttachmentsTotalMbps = 425_090_419

InterconnectTotalGbps = 285_341_866

InternalAddresses = 197_899_392

InternalTrafficDirectorForwardingRules = 266_433_668

InPlaceSnapshots = 151_359_133

InUseAddresses = 402_125_072

InUseBackupSchedules = 32_786_705

InUseSnapshotSchedules = 462_104_083

LocalSsdTotalGb = 330_878_021

M1Cpus = 37_203_366

M2Cpus = 65_832_517

M3Cpus = 94_461_668

MachineImages = 446_986_640

N2aCpus = 265_855_917

N2dCpus = 351_743_370

N2Cpus = 416_465_286

Networks = 485_481_477

NetworkAttachments = 149_028_575

NetworkEndpointGroups = 102_144_909

NetworkFirewallPolicies = 101_117_374

NetLbSecurityPoliciesPerRegion = 157_892_269

NetLbSecurityPolicyRulesPerRegion = 356_090_931

NetLbSecurityPolicyRuleAttributesPerRegion = 311_243_888

NodeGroups = 24_624_817

NodeTemplates = 474_896_668

NvidiaA10080gbGpus = 286_389_320

NvidiaA100Gpus = 504_872_978

NvidiaK80Gpus = 163_886_599

NvidiaL4Gpus = 491_923_130

NvidiaP100Gpus = 236_601_633

NvidiaP100VwsGpus = 213_970_574

NvidiaP4Gpus = 283_841_470

NvidiaP4VwsGpus = 528_296_619

NvidiaT4Gpus = 75_759_810

NvidiaT4VwsGpus = 319_813_039

NvidiaV100Gpus = 129_293_095

PacketMirrorings = 15_578_407

PdExtremeTotalProvisionedIops = 69_593_965

PreemptibleCpus = 251_184_841

PreemptibleLocalSsdGb = 260_819_336

PreemptibleNvidiaA10080gbGpus = 151_942_410

PreemptibleNvidiaA100Gpus = 68_832_784

PreemptibleNvidiaH100Gpus = 301_553_431

PreemptibleNvidiaK80Gpus = 374_960_201

PreemptibleNvidiaL4Gpus = 100_408_376

PreemptibleNvidiaP100Gpus = 337_432_351

PreemptibleNvidiaP100VwsGpus = 313_544_076

PreemptibleNvidiaP4Gpus = 429_197_628

PreemptibleNvidiaP4VwsGpus = 252_981_545

PreemptibleNvidiaT4Gpus = 221_115_968

PreemptibleNvidiaT4VwsGpus = 44_497_965

PreemptibleNvidiaV100Gpus = 230_123_813

PreemptibleTpuLiteDeviceV5 = 174_262_523

PreemptibleTpuLitePodsliceV5 = 112_281_796

PreemptibleTpuPodsliceV4 = 425_541_132

PscIlbConsumerForwardingRulesPerProducerNetwork = 231_164_291

PscInternalLbForwardingRules = 169_005_435

PublicAdvertisedPrefixes = 471_371_980

PublicDelegatedPrefixes = 532_465_974

RegionalAutoscalers = 29_363_772

RegionalExternalManagedBackendServices = 4_240_989

RegionalExternalNetworkLbBackendServices = 409_564_525

RegionalInstanceGroupManagers = 37_543_696

RegionalInternalLbBackendServices = 137_983_760

RegionalInternalManagedBackendServices = 96_282_539

RegionalInternalTrafficDirectorBackendServices = 483_162_968

Reservations = 32_644_647

ResourcePolicies = 83_955_297

Routers = 493_018_666

Routes = 275_680_074

SecurityPolicies = 189_518_703

SecurityPoliciesPerRegion = 249_041_734

SecurityPolicyAdvancedRulesPerRegion = 371_815_341

SecurityPolicyCevalRules = 470_815_689

SecurityPolicyRules = 203_549_225

SecurityPolicyRulesPerRegion = 126_510_156

ServiceAttachments = 471_521_510

Snapshots = 343_405_327

The total number of snapshots allowed for a single project.

SsdTotalGb = 161_732_561

SslCertificates = 378_372_399

SslPolicies = 523_254_339

StaticAddresses = 93_624_049

StaticByoipAddresses = 275_809_649

StaticExternalIpv6AddressRanges = 472_346_774

Subnetworks = 421_330_469

T2aCpus = 522_170_599

T2dCpus = 71_187_140

TargetHttpsProxies = 219_522_506

TargetHttpProxies = 164_117_155

TargetInstances = 284_519_728

TargetPools = 348_261_257

TargetSslProxies = 159_216_235

TargetTcpProxies = 182_243_136

TargetVpnGateways = 75_029_928

TpuLiteDeviceV5 = 449_577_597

TpuLitePodsliceV5 = 12_708_294

TpuPodsliceV4 = 214_467_530

UrlMaps = 378_660_743

VariableIpv6PublicDelegatedPrefixes = 128_400_161

VpnGateways = 35_620_282

VpnTunnels = 104_327_296

XpnServiceProjects = 95_191_981

Implementations

impl Metric

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Metric.

pub fn from_i32(value: i32) -> Option<Metric>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Metric, or None if value is not a valid variant.

impl Metric

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.