Struct google_api_proto::google::cloud::compute::v1::Backend

pub struct Backend {

    pub balancing_mode: Option<String>,
    pub capacity_scaler: Option<f32>,
    pub description: Option<String>,
    pub failover: Option<bool>,
    pub group: Option<String>,
    pub max_connections: Option<i32>,
    pub max_connections_per_endpoint: Option<i32>,
    pub max_connections_per_instance: Option<i32>,
    pub max_rate: Option<i32>,
    pub max_rate_per_endpoint: Option<f32>,
    pub max_rate_per_instance: Option<f32>,
    pub max_utilization: Option<f32>,
    pub preference: Option<String>,

}

Message containing information of one individual backend.

Fields
balancing_mode: Option<String>

Specifies how to determine whether the backend of a load balancer can handle additional traffic or is fully loaded. For usage guidelines, see Connection balancing mode. Backends must use compatible balancing modes. For more information, see Supported balancing modes and target capacity settings and Restrictions and guidance for instance groups. Note: Currently, if you use the API to configure incompatible balancing modes, the configuration might be accepted even though it has no impact and is ignored. Specifically, Backend.maxUtilization is ignored when Backend.balancingMode is RATE. In the future, this incompatible combination will be rejected. Check the BalancingMode enum for the list of possible values.

capacity_scaler: Option<f32>

A multiplier applied to the backend’s target capacity of its balancing mode. The default value is 1, which means the group serves up to 100% of its configured capacity (depending on balancingMode). A setting of 0 means the group is completely drained, offering 0% of its available capacity. The valid ranges are 0.0 and [0.1,1.0]. You cannot configure a setting larger than 0 and smaller than 0.1. You cannot configure a setting of 0 when there is only one backend attached to the backend service. Not available with backends that don’t support using a balancingMode. This includes backends such as global internet NEGs, regional serverless NEGs, and PSC NEGs.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

failover: Option<bool>

This field designates whether this is a failover backend. More than one failover backend can be configured for a given BackendService.

group: Option<String>

The fully-qualified URL of an instance group or network endpoint group (NEG) resource. To determine what types of backends a load balancer supports, see the Backend services overview. You must use the fully-qualified URL (starting with https://www.googleapis.com/) to specify the instance group or NEG. Partial URLs are not supported.

max_connections: Option<i32>

Defines a target maximum number of simultaneous connections. For usage guidelines, see Connection balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is RATE.

max_connections_per_endpoint: Option<i32>

Defines a target maximum number of simultaneous connections. For usage guidelines, see Connection balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is RATE.

max_connections_per_instance: Option<i32>

Defines a target maximum number of simultaneous connections. For usage guidelines, see Connection balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is RATE.

max_rate: Option<i32>

Defines a maximum number of HTTP requests per second (RPS). For usage guidelines, see Rate balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is CONNECTION.

max_rate_per_endpoint: Option<f32>

Defines a maximum target for requests per second (RPS). For usage guidelines, see Rate balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is CONNECTION.

max_rate_per_instance: Option<f32>

Defines a maximum target for requests per second (RPS). For usage guidelines, see Rate balancing mode and Utilization balancing mode. Not available if the backend’s balancingMode is CONNECTION.

max_utilization: Option<f32>

Optional parameter to define a target capacity for the UTILIZATION balancing mode. The valid range is [0.0, 1.0]. For usage guidelines, see Utilization balancing mode.

preference: Option<String>

This field indicates whether this backend should be fully utilized before sending traffic to backends with default preference. The possible values are: - PREFERRED: Backends with this preference level will be filled up to their capacity limits first, based on RTT. - DEFAULT: If preferred backends don’t have enough capacity, backends in this layer would be used and traffic would be assigned based on the load balancing algorithm you use. This is the default Check the Preference enum for the list of possible values.

Implementations

impl Backend

pub fn max_rate_per_instance(&self) -> f32

Returns the value of max_rate_per_instance, or the default value if max_rate_per_instance is unset.

pub fn group(&self) -> &str

Returns the value of group, or the default value if group is unset.

pub fn max_connections_per_instance(&self) -> i32

Returns the value of max_connections_per_instance, or the default value if max_connections_per_instance is unset.

pub fn max_connections(&self) -> i32

Returns the value of max_connections, or the default value if max_connections is unset.

pub fn max_rate_per_endpoint(&self) -> f32

Returns the value of max_rate_per_endpoint, or the default value if max_rate_per_endpoint is unset.

pub fn failover(&self) -> bool

Returns the value of failover, or the default value if failover is unset.

pub fn max_utilization(&self) -> f32

Returns the value of max_utilization, or the default value if max_utilization is unset.

pub fn preference(&self) -> &str

Returns the value of preference, or the default value if preference is unset.

pub fn max_connections_per_endpoint(&self) -> i32

Returns the value of max_connections_per_endpoint, or the default value if max_connections_per_endpoint is unset.

pub fn capacity_scaler(&self) -> f32

Returns the value of capacity_scaler, or the default value if capacity_scaler is unset.

pub fn max_rate(&self) -> i32

Returns the value of max_rate, or the default value if max_rate is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn balancing_mode(&self) -> &str

Returns the value of balancing_mode, or the default value if balancing_mode is unset.