Struct google_api_proto::google::cloud::compute::v1::BfdPacket

pub struct BfdPacket {

    pub authentication_present: Option<bool>,
    pub control_plane_independent: Option<bool>,
    pub demand: Option<bool>,
    pub diagnostic: Option<String>,
    pub final: Option<bool>,
    pub length: Option<u32>,
    pub min_echo_rx_interval_ms: Option<u32>,
    pub min_rx_interval_ms: Option<u32>,
    pub min_tx_interval_ms: Option<u32>,
    pub multiplier: Option<u32>,
    pub multipoint: Option<bool>,
    pub my_discriminator: Option<u32>,
    pub poll: Option<bool>,
    pub state: Option<String>,
    pub version: Option<u32>,
    pub your_discriminator: Option<u32>,

}

Fields
authentication_present: Option<bool>

The Authentication Present bit of the BFD packet. This is specified in section 4.1 of RFC5880

control_plane_independent: Option<bool>

The Control Plane Independent bit of the BFD packet. This is specified in section 4.1 of RFC5880

demand: Option<bool>

The demand bit of the BFD packet. This is specified in section 4.1 of RFC5880

diagnostic: Option<String>

The diagnostic code specifies the local system’s reason for the last change in session state. This allows remote systems to determine the reason that the previous session failed, for example. These diagnostic codes are specified in section 4.1 of RFC5880 Check the Diagnostic enum for the list of possible values.

final: Option<bool>

The Final bit of the BFD packet. This is specified in section 4.1 of RFC5880

length: Option<u32>

The length of the BFD Control packet in bytes. This is specified in section 4.1 of RFC5880

min_echo_rx_interval_ms: Option<u32>

The Required Min Echo RX Interval value in the BFD packet. This is specified in section 4.1 of RFC5880

min_rx_interval_ms: Option<u32>

The Required Min RX Interval value in the BFD packet. This is specified in section 4.1 of RFC5880

min_tx_interval_ms: Option<u32>

The Desired Min TX Interval value in the BFD packet. This is specified in section 4.1 of RFC5880

multiplier: Option<u32>

The detection time multiplier of the BFD packet. This is specified in section 4.1 of RFC5880

multipoint: Option<bool>

The multipoint bit of the BFD packet. This is specified in section 4.1 of RFC5880

my_discriminator: Option<u32>

The My Discriminator value in the BFD packet. This is specified in section 4.1 of RFC5880

poll: Option<bool>

The Poll bit of the BFD packet. This is specified in section 4.1 of RFC5880

state: Option<String>

The current BFD session state as seen by the transmitting system. These states are specified in section 4.1 of RFC5880 Check the State enum for the list of possible values.

version: Option<u32>

The version number of the BFD protocol, as specified in section 4.1 of RFC5880.

your_discriminator: Option<u32>

The Your Discriminator value in the BFD packet. This is specified in section 4.1 of RFC5880

Implementations

impl BfdPacket

pub fn poll(&self) -> bool

Returns the value of poll, or the default value if poll is unset.

pub fn control_plane_independent(&self) -> bool

Returns the value of control_plane_independent, or the default value if control_plane_independent is unset.

pub fn diagnostic(&self) -> &str

Returns the value of diagnostic, or the default value if diagnostic is unset.

pub fn my_discriminator(&self) -> u32

Returns the value of my_discriminator, or the default value if my_discriminator is unset.

pub fn min_echo_rx_interval_ms(&self) -> u32

Returns the value of min_echo_rx_interval_ms, or the default value if min_echo_rx_interval_ms is unset.

pub fn final(&self) -> bool

Returns the value of final, or the default value if final is unset.

pub fn authentication_present(&self) -> bool

Returns the value of authentication_present, or the default value if authentication_present is unset.

pub fn state(&self) -> &str

Returns the value of state, or the default value if state is unset.

pub fn multiplier(&self) -> u32

Returns the value of multiplier, or the default value if multiplier is unset.

pub fn multipoint(&self) -> bool

Returns the value of multipoint, or the default value if multipoint is unset.

pub fn demand(&self) -> bool

Returns the value of demand, or the default value if demand is unset.

pub fn version(&self) -> u32

Returns the value of version, or the default value if version is unset.

pub fn min_rx_interval_ms(&self) -> u32

Returns the value of min_rx_interval_ms, or the default value if min_rx_interval_ms is unset.

pub fn length(&self) -> u32

Returns the value of length, or the default value if length is unset.

pub fn your_discriminator(&self) -> u32

Returns the value of your_discriminator, or the default value if your_discriminator is unset.

pub fn min_tx_interval_ms(&self) -> u32

Returns the value of min_tx_interval_ms, or the default value if min_tx_interval_ms is unset.