Struct google_api_proto::google::cloud::compute::v1::Image

pub struct Image {

    pub architecture: Option<String>,
    pub archive_size_bytes: Option<i64>,
    pub creation_timestamp: Option<String>,
    pub deprecated: Option<DeprecationStatus>,
    pub description: Option<String>,
    pub disk_size_gb: Option<i64>,
    pub family: Option<String>,
    pub guest_os_features: Vec<GuestOsFeature>,
    pub id: Option<u64>,
    pub image_encryption_key: Option<CustomerEncryptionKey>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub license_codes: Vec<i64>,
    pub licenses: Vec<String>,
    pub name: Option<String>,
    pub raw_disk: Option<RawDisk>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub shielded_instance_initial_state: Option<InitialStateConfig>,
    pub source_disk: Option<String>,
    pub source_disk_encryption_key: Option<CustomerEncryptionKey>,
    pub source_disk_id: Option<String>,
    pub source_image: Option<String>,
    pub source_image_encryption_key: Option<CustomerEncryptionKey>,
    pub source_image_id: Option<String>,
    pub source_snapshot: Option<String>,
    pub source_snapshot_encryption_key: Option<CustomerEncryptionKey>,
    pub source_snapshot_id: Option<String>,
    pub source_type: Option<String>,
    pub status: Option<String>,
    pub storage_locations: Vec<String>,

}

Represents an Image resource. You can use images to create boot disks for your VM instances. For more information, read Images.

Fields
architecture: Option<String>

The architecture of the image. Valid values are ARM64 or X86_64. Check the Architecture enum for the list of possible values.

archive_size_bytes: Option<i64>

Size of the image tar.gz archive stored in Google Cloud Storage (in bytes).

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

deprecated: Option<DeprecationStatus>

The deprecation status associated with this image.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disk_size_gb: Option<i64>

Size of the image when restored onto a persistent disk (in GB).

family: Option<String>

The name of the image family to which this image belongs. The image family name can be from a publicly managed image family provided by Compute Engine, or from a custom image family you create. For example, centos-stream-9 is a publicly available image family. For more information, see Image family best practices. When creating disks, you can specify an image family instead of a specific image name. The image family always returns its latest image that is not deprecated. The name of the image family must comply with RFC1035.

guest_os_features: Vec<GuestOsFeature>

A list of features to enable on the guest operating system. Applicable only for bootable images. To see a list of available options, see the guestOSfeatures[].type parameter.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

image_encryption_key: Option<CustomerEncryptionKey>

Encrypts the image using a customer-supplied encryption key. After you encrypt an image with a customer-supplied key, you must provide the same key if you use the image later (e.g. to create a disk from the image). Customer-supplied encryption keys do not protect access to metadata of the disk. If you do not provide an encryption key when creating the image, then the disk will be encrypted using an automatically generated key and you do not need to provide a key to use the image later.

kind: Option<String>

[Output Only] Type of the resource. Always compute#image for images.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this image, which is essentially a hash of the labels used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve an image.

labels: BTreeMap<String, String>

Labels to apply to this image. These can be later modified by the setLabels method.

license_codes: Vec<i64>

Integer license codes indicating which licenses are attached to this image.

licenses: Vec<String>

Any applicable license URI.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

raw_disk: Option<RawDisk>

The parameters of the raw disk image.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

shielded_instance_initial_state: Option<InitialStateConfig>

Set the secure boot keys of shielded instance.

source_disk: Option<String>

URL of the source disk used to create this image. For example, the following are valid values: - https://www.googleapis.com/compute/v1/projects/project/zones/zone /disks/disk - projects/project/zones/zone/disks/disk - zones/zone/disks/disk In order to create an image, you must provide the full or partial URL of one of the following: - The rawDisk.source URL - The sourceDisk URL - The sourceImage URL - The sourceSnapshot URL

source_disk_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source disk. Required if the source disk is protected by a customer-supplied encryption key.

source_disk_id: Option<String>

[Output Only] The ID value of the disk used to create this image. This value may be used to determine whether the image was taken from the current or a previous instance of a given disk name.

source_image: Option<String>

URL of the source image used to create this image. The following are valid formats for the URL: - https://www.googleapis.com/compute/v1/projects/project_id/global/ images/image_name - projects/project_id/global/images/image_name In order to create an image, you must provide the full or partial URL of one of the following: - The rawDisk.source URL - The sourceDisk URL - The sourceImage URL - The sourceSnapshot URL

source_image_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source image. Required if the source image is protected by a customer-supplied encryption key.

source_image_id: Option<String>

[Output Only] The ID value of the image used to create this image. This value may be used to determine whether the image was taken from the current or a previous instance of a given image name.

source_snapshot: Option<String>

URL of the source snapshot used to create this image. The following are valid formats for the URL: - https://www.googleapis.com/compute/v1/projects/project_id/global/ snapshots/snapshot_name - projects/project_id/global/snapshots/snapshot_name In order to create an image, you must provide the full or partial URL of one of the following: - The rawDisk.source URL - The sourceDisk URL - The sourceImage URL - The sourceSnapshot URL

source_snapshot_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source snapshot. Required if the source snapshot is protected by a customer-supplied encryption key.

source_snapshot_id: Option<String>

[Output Only] The ID value of the snapshot used to create this image. This value may be used to determine whether the snapshot was taken from the current or a previous instance of a given snapshot name.

source_type: Option<String>

The type of the image used to create this disk. The default and only valid value is RAW. Check the SourceType enum for the list of possible values.

status: Option<String>

[Output Only] The status of the image. An image can be used to create other resources, such as instances, only after the image has been successfully created and the status is set to READY. Possible values are FAILED, PENDING, or READY. Check the Status enum for the list of possible values.

storage_locations: Vec<String>

Cloud Storage bucket storage location of the image (regional or multi-regional).

Implementations

impl Image

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn source_image(&self) -> &str

Returns the value of source_image, or the default value if source_image is unset.

pub fn source_image_id(&self) -> &str

Returns the value of source_image_id, or the default value if source_image_id is unset.

pub fn source_snapshot_id(&self) -> &str

Returns the value of source_snapshot_id, or the default value if source_snapshot_id is unset.

pub fn source_snapshot(&self) -> &str

Returns the value of source_snapshot, or the default value if source_snapshot is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn architecture(&self) -> &str

Returns the value of architecture, or the default value if architecture is unset.

pub fn disk_size_gb(&self) -> i64

Returns the value of disk_size_gb, or the default value if disk_size_gb is unset.

pub fn family(&self) -> &str

Returns the value of family, or the default value if family is unset.

pub fn archive_size_bytes(&self) -> i64

Returns the value of archive_size_bytes, or the default value if archive_size_bytes is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn source_disk(&self) -> &str

Returns the value of source_disk, or the default value if source_disk is unset.

pub fn source_type(&self) -> &str

Returns the value of source_type, or the default value if source_type is unset.

pub fn source_disk_id(&self) -> &str

Returns the value of source_disk_id, or the default value if source_disk_id is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.