Struct google_api_proto::google::cloud::compute::v1::Instance

pub struct Instance {

    pub advanced_machine_features: Option<AdvancedMachineFeatures>,
    pub can_ip_forward: Option<bool>,
    pub confidential_instance_config: Option<ConfidentialInstanceConfig>,
    pub cpu_platform: Option<String>,
    pub creation_timestamp: Option<String>,
    pub deletion_protection: Option<bool>,
    pub description: Option<String>,
    pub disks: Vec<AttachedDisk>,
    pub display_device: Option<DisplayDevice>,
    pub fingerprint: Option<String>,
    pub guest_accelerators: Vec<AcceleratorConfig>,
    pub hostname: Option<String>,
    pub id: Option<u64>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub last_start_timestamp: Option<String>,
    pub last_stop_timestamp: Option<String>,
    pub last_suspended_timestamp: Option<String>,
    pub machine_type: Option<String>,
    pub metadata: Option<Metadata>,
    pub min_cpu_platform: Option<String>,
    pub name: Option<String>,
    pub network_interfaces: Vec<NetworkInterface>,
    pub network_performance_config: Option<NetworkPerformanceConfig>,
    pub params: Option<InstanceParams>,
    pub private_ipv6_google_access: Option<String>,
    pub reservation_affinity: Option<ReservationAffinity>,
    pub resource_policies: Vec<String>,
    pub satisfies_pzs: Option<bool>,
    pub scheduling: Option<Scheduling>,
    pub self_link: Option<String>,
    pub service_accounts: Vec<ServiceAccount>,
    pub shielded_instance_config: Option<ShieldedInstanceConfig>,
    pub shielded_instance_integrity_policy: Option<ShieldedInstanceIntegrityPolicy>,
    pub source_machine_image: Option<String>,
    pub source_machine_image_encryption_key: Option<CustomerEncryptionKey>,
    pub start_restricted: Option<bool>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub tags: Option<Tags>,
    pub zone: Option<String>,

}

Represents an Instance resource. An instance is a virtual machine that is hosted on Google Cloud Platform. For more information, read Virtual Machine Instances.

Fields
advanced_machine_features: Option<AdvancedMachineFeatures>

Controls for advanced machine-related behavior features.

can_ip_forward: Option<bool>

Allows this instance to send and receive packets with non-matching destination or source IPs. This is required if you plan to use this instance to forward routes. For more information, see Enabling IP Forwarding .

confidential_instance_config: Option<ConfidentialInstanceConfig>cpu_platform: Option<String>

[Output Only] The CPU platform used by this instance.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

deletion_protection: Option<bool>

Whether the resource should be protected against deletion.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disks: Vec<AttachedDisk>

Array of disks associated with this instance. Persistent disks must be created before you can assign them.

display_device: Option<DisplayDevice>

Enables display device for the instance.

fingerprint: Option<String>

Specifies a fingerprint for this resource, which is essentially a hash of the instance’s contents and used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update the instance. You must always provide an up-to-date fingerprint hash in order to update the instance. To see the latest fingerprint, make get() request to the instance.

guest_accelerators: Vec<AcceleratorConfig>

A list of the type and count of accelerator cards attached to the instance.

hostname: Option<String>

Specifies the hostname of the instance. The specified hostname must be RFC1035 compliant. If hostname is not specified, the default hostname is [INSTANCE_NAME].c.[PROJECT_ID].internal when using the global DNS, and [INSTANCE_NAME].[ZONE].c.[PROJECT_ID].internal when using zonal DNS.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#instance for instances.

label_fingerprint: Option<String>

A fingerprint for this request, which is essentially a hash of the label’s contents and used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels. To see the latest fingerprint, make get() request to the instance.

labels: BTreeMap<String, String>

Labels to apply to this instance. These can be later modified by the setLabels method.

last_start_timestamp: Option<String>

[Output Only] Last start timestamp in RFC3339 text format.

last_stop_timestamp: Option<String>

[Output Only] Last stop timestamp in RFC3339 text format.

last_suspended_timestamp: Option<String>

[Output Only] Last suspended timestamp in RFC3339 text format.

machine_type: Option<String>

Full or partial URL of the machine type resource to use for this instance, in the format: zones/zone/machineTypes/machine-type. This is provided by the client when the instance is created. For example, the following is a valid partial url to a predefined machine type: zones/us-central1-f/machineTypes/n1-standard-1 To create a custom machine type, provide a URL to a machine type in the following format, where CPUS is 1 or an even number up to 32 (2, 4, 6, … 24, etc), and MEMORY is the total memory for this instance. Memory must be a multiple of 256 MB and must be supplied in MB (e.g. 5 GB of memory is 5120 MB): zones/zone/machineTypes/custom-CPUS-MEMORY For example: zones/us-central1-f/machineTypes/custom-4-5120 For a full list of restrictions, read the Specifications for custom machine types.

metadata: Option<Metadata>

The metadata key/value pairs assigned to this instance. This includes custom metadata and predefined keys.

min_cpu_platform: Option<String>

Specifies a minimum CPU platform for the VM instance. Applicable values are the friendly names of CPU platforms, such as minCpuPlatform: “Intel Haswell” or minCpuPlatform: “Intel Sandy Bridge”.

name: Option<String>

The name of the resource, provided by the client when initially creating the resource. The resource name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression \[a-z]([-a-z0-9]*[a-z0-9\])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network_interfaces: Vec<NetworkInterface>

An array of network configurations for this instance. These specify how interfaces are configured to interact with other network services, such as connecting to the internet. Multiple interfaces are supported per instance.

network_performance_config: Option<NetworkPerformanceConfig>params: Option<InstanceParams>

Input only. [Input Only] Additional params passed with the request, but not persisted as part of resource payload.

private_ipv6_google_access: Option<String>

The private IPv6 google access type for the VM. If not specified, use INHERIT_FROM_SUBNETWORK as default. Check the PrivateIpv6GoogleAccess enum for the list of possible values.

reservation_affinity: Option<ReservationAffinity>

Specifies the reservations that this instance can consume from.

resource_policies: Vec<String>

Resource policies applied to this instance.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

scheduling: Option<Scheduling>

Sets the scheduling options for this instance.

self_link: Option<String>

[Output Only] Server-defined URL for this resource.

service_accounts: Vec<ServiceAccount>

A list of service accounts, with their specified scopes, authorized for this instance. Only one service account per VM instance is supported. Service accounts generate access tokens that can be accessed through the metadata server and used to authenticate applications on the instance. See Service Accounts for more information.

shielded_instance_config: Option<ShieldedInstanceConfig>shielded_instance_integrity_policy: Option<ShieldedInstanceIntegrityPolicy>source_machine_image: Option<String>

Source machine image

source_machine_image_encryption_key: Option<CustomerEncryptionKey>

Source machine image encryption key when creating an instance from a machine image.

start_restricted: Option<bool>

[Output Only] Whether a VM has been restricted for start because Compute Engine has detected suspicious activity.

status: Option<String>

[Output Only] The status of the instance. One of the following values: PROVISIONING, STAGING, RUNNING, STOPPING, SUSPENDING, SUSPENDED, REPAIRING, and TERMINATED. For more information about the status of the instance, see Instance life cycle. Check the Status enum for the list of possible values.

status_message: Option<String>

[Output Only] An optional, human-readable explanation of the status.

tags: Option<Tags>

Tags to apply to this instance. Tags are used to identify valid sources or targets for network firewalls and are specified by the client during instance creation. The tags can be later modified by the setTags method. Each tag within the list must comply with RFC1035. Multiple tags can be specified via the ‘tags.items’ field.

zone: Option<String>

[Output Only] URL of the zone where the instance resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

Implementations

impl Instance

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn zone(&self) -> &str

Returns the value of zone, or the default value if zone is unset.

pub fn source_machine_image(&self) -> &str

Returns the value of source_machine_image, or the default value if source_machine_image is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn private_ipv6_google_access(&self) -> &str

Returns the value of private_ipv6_google_access, or the default value if private_ipv6_google_access is unset.

pub fn start_restricted(&self) -> bool

Returns the value of start_restricted, or the default value if start_restricted is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn machine_type(&self) -> &str

Returns the value of machine_type, or the default value if machine_type is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn hostname(&self) -> &str

Returns the value of hostname, or the default value if hostname is unset.

pub fn min_cpu_platform(&self) -> &str

Returns the value of min_cpu_platform, or the default value if min_cpu_platform is unset.

pub fn status_message(&self) -> &str

Returns the value of status_message, or the default value if status_message is unset.

pub fn last_suspended_timestamp(&self) -> &str

Returns the value of last_suspended_timestamp, or the default value if last_suspended_timestamp is unset.

pub fn cpu_platform(&self) -> &str

Returns the value of cpu_platform, or the default value if cpu_platform is unset.

pub fn last_stop_timestamp(&self) -> &str

Returns the value of last_stop_timestamp, or the default value if last_stop_timestamp is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn last_start_timestamp(&self) -> &str

Returns the value of last_start_timestamp, or the default value if last_start_timestamp is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn deletion_protection(&self) -> bool

Returns the value of deletion_protection, or the default value if deletion_protection is unset.

pub fn can_ip_forward(&self) -> bool

Returns the value of can_ip_forward, or the default value if can_ip_forward is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.