Struct google_api_proto::google::cloud::compute::v1::InstanceGroupManager

pub struct InstanceGroupManager {

    pub all_instances_config: Option<InstanceGroupManagerAllInstancesConfig>,
    pub auto_healing_policies: Vec<InstanceGroupManagerAutoHealingPolicy>,
    pub base_instance_name: Option<String>,
    pub creation_timestamp: Option<String>,
    pub current_actions: Option<InstanceGroupManagerActionsSummary>,
    pub description: Option<String>,
    pub distribution_policy: Option<DistributionPolicy>,
    pub fingerprint: Option<String>,
    pub id: Option<u64>,
    pub instance_group: Option<String>,
    pub instance_lifecycle_policy: Option<InstanceGroupManagerInstanceLifecyclePolicy>,
    pub instance_template: Option<String>,
    pub kind: Option<String>,
    pub list_managed_instances_results: Option<String>,
    pub name: Option<String>,
    pub named_ports: Vec<NamedPort>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub stateful_policy: Option<StatefulPolicy>,
    pub status: Option<InstanceGroupManagerStatus>,
    pub target_pools: Vec<String>,
    pub target_size: Option<i32>,
    pub update_policy: Option<InstanceGroupManagerUpdatePolicy>,
    pub versions: Vec<InstanceGroupManagerVersion>,
    pub zone: Option<String>,

}

Represents a Managed Instance Group resource. An instance group is a collection of VM instances that you can manage as a single entity. For more information, read Instance groups. For zonal Managed Instance Group, use the instanceGroupManagers resource. For regional Managed Instance Group, use the regionInstanceGroupManagers resource.

Fields
all_instances_config: Option<InstanceGroupManagerAllInstancesConfig>

Specifies configuration that overrides the instance template configuration for the group.

auto_healing_policies: Vec<InstanceGroupManagerAutoHealingPolicy>

The autohealing policy for this managed instance group. You can specify only one value.

base_instance_name: Option<String>

The base instance name to use for instances in this group. The value must be 1-58 characters long. Instances are named by appending a hyphen and a random four-character string to the base instance name. The base instance name must comply with RFC1035.

creation_timestamp: Option<String>

[Output Only] The creation timestamp for this managed instance group in RFC3339 text format.

current_actions: Option<InstanceGroupManagerActionsSummary>

[Output Only] The list of instance actions and the number of instances in this managed instance group that are scheduled for each of those actions.

description: Option<String>

An optional description of this resource.

distribution_policy: Option<DistributionPolicy>

Policy specifying the intended distribution of managed instances across zones in a regional managed instance group.

fingerprint: Option<String>

Fingerprint of this resource. This field may be used in optimistic locking. It will be ignored when inserting an InstanceGroupManager. An up-to-date fingerprint must be provided in order to update the InstanceGroupManager, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve an InstanceGroupManager.

id: Option<u64>

[Output Only] A unique identifier for this resource type. The server generates this identifier.

instance_group: Option<String>

[Output Only] The URL of the Instance Group resource.

instance_lifecycle_policy: Option<InstanceGroupManagerInstanceLifecyclePolicy>

The repair policy for this managed instance group.

instance_template: Option<String>

The URL of the instance template that is specified for this managed instance group. The group uses this template to create all new instances in the managed instance group. The templates for existing instances in the group do not change unless you run recreateInstances, run applyUpdatesToInstances, or set the group’s updatePolicy.type to PROACTIVE.

kind: Option<String>

[Output Only] The resource type, which is always compute#instanceGroupManager for managed instance groups.

list_managed_instances_results: Option<String>

Pagination behavior of the listManagedInstances API method for this managed instance group. Check the ListManagedInstancesResults enum for the list of possible values.

name: Option<String>

The name of the managed instance group. The name must be 1-63 characters long, and comply with RFC1035.

named_ports: Vec<NamedPort>

Named ports configured for the Instance Groups complementary to this Instance Group Manager.

region: Option<String>

[Output Only] The URL of the region where the managed instance group resides (for regional resources).

self_link: Option<String>

[Output Only] The URL for this managed instance group. The server defines this URL.

stateful_policy: Option<StatefulPolicy>

Stateful configuration for this Instanced Group Manager

status: Option<InstanceGroupManagerStatus>

[Output Only] The status of this managed instance group.

target_pools: Vec<String>

The URLs for all TargetPool resources to which instances in the instanceGroup field are added. The target pools automatically apply to all of the instances in the managed instance group.

target_size: Option<i32>

The target number of running instances for this managed instance group. You can reduce this number by using the instanceGroupManager deleteInstances or abandonInstances methods. Resizing the group also changes this number.

update_policy: Option<InstanceGroupManagerUpdatePolicy>

The update policy for this managed instance group.

versions: Vec<InstanceGroupManagerVersion>

Specifies the instance templates used by this managed instance group to create instances. Each version is defined by an instanceTemplate and a name. Every version can appear at most once per instance group. This field overrides the top-level instanceTemplate field. Read more about the relationships between these fields. Exactly one version must leave the targetSize field unset. That version will be applied to all remaining instances. For more information, read about canary updates.

zone: Option<String>

[Output Only] The URL of a zone where the managed instance group is located (for zonal resources).

Implementations

impl InstanceGroupManager

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn zone(&self) -> &str

Returns the value of zone, or the default value if zone is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn target_size(&self) -> i32

Returns the value of target_size, or the default value if target_size is unset.

pub fn instance_group(&self) -> &str

Returns the value of instance_group, or the default value if instance_group is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn list_managed_instances_results(&self) -> &str

Returns the value of list_managed_instances_results, or the default value if list_managed_instances_results is unset.

pub fn instance_template(&self) -> &str

Returns the value of instance_template, or the default value if instance_template is unset.

pub fn base_instance_name(&self) -> &str

Returns the value of base_instance_name, or the default value if base_instance_name is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.