Struct google_api_proto::google::cloud::compute::v1::InstanceProperties

pub struct InstanceProperties {

    pub advanced_machine_features: Option<AdvancedMachineFeatures>,
    pub can_ip_forward: Option<bool>,
    pub confidential_instance_config: Option<ConfidentialInstanceConfig>,
    pub description: Option<String>,
    pub disks: Vec<AttachedDisk>,
    pub guest_accelerators: Vec<AcceleratorConfig>,
    pub key_revocation_action_type: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub machine_type: Option<String>,
    pub metadata: Option<Metadata>,
    pub min_cpu_platform: Option<String>,
    pub network_interfaces: Vec<NetworkInterface>,
    pub network_performance_config: Option<NetworkPerformanceConfig>,
    pub private_ipv6_google_access: Option<String>,
    pub reservation_affinity: Option<ReservationAffinity>,
    pub resource_manager_tags: BTreeMap<String, String>,
    pub resource_policies: Vec<String>,
    pub scheduling: Option<Scheduling>,
    pub service_accounts: Vec<ServiceAccount>,
    pub shielded_instance_config: Option<ShieldedInstanceConfig>,
    pub tags: Option<Tags>,

}

Fields
advanced_machine_features: Option<AdvancedMachineFeatures>

Controls for advanced machine-related behavior features. Note that for MachineImage, this is not supported yet.

can_ip_forward: Option<bool>

Enables instances created based on these properties to send packets with source IP addresses other than their own and receive packets with destination IP addresses other than their own. If these instances will be used as an IP gateway or it will be set as the next-hop in a Route resource, specify true. If unsure, leave this set to false. See the Enable IP forwarding documentation for more information.

confidential_instance_config: Option<ConfidentialInstanceConfig>

Specifies the Confidential Instance options. Note that for MachineImage, this is not supported yet.

description: Option<String>

An optional text description for the instances that are created from these properties.

disks: Vec<AttachedDisk>

An array of disks that are associated with the instances that are created from these properties.

guest_accelerators: Vec<AcceleratorConfig>

A list of guest accelerator cards’ type and count to use for instances created from these properties.

key_revocation_action_type: Option<String>

KeyRevocationActionType of the instance. Supported options are “STOP” and “NONE”. The default value is “NONE” if it is not specified. Check the KeyRevocationActionType enum for the list of possible values.

labels: BTreeMap<String, String>

Labels to apply to instances that are created from these properties.

machine_type: Option<String>

The machine type to use for instances that are created from these properties.

metadata: Option<Metadata>

The metadata key/value pairs to assign to instances that are created from these properties. These pairs can consist of custom metadata or predefined keys. See Project and instance metadata for more information.

min_cpu_platform: Option<String>

Minimum cpu/platform to be used by instances. The instance may be scheduled on the specified or newer cpu/platform. Applicable values are the friendly names of CPU platforms, such as minCpuPlatform: “Intel Haswell” or minCpuPlatform: “Intel Sandy Bridge”. For more information, read Specifying a Minimum CPU Platform.

network_interfaces: Vec<NetworkInterface>

An array of network access configurations for this interface.

network_performance_config: Option<NetworkPerformanceConfig>

Note that for MachineImage, this is not supported yet.

private_ipv6_google_access: Option<String>

The private IPv6 google access type for VMs. If not specified, use INHERIT_FROM_SUBNETWORK as default. Note that for MachineImage, this is not supported yet. Check the PrivateIpv6GoogleAccess enum for the list of possible values.

reservation_affinity: Option<ReservationAffinity>

Specifies the reservations that instances can consume from. Note that for MachineImage, this is not supported yet.

resource_manager_tags: BTreeMap<String, String>

Resource manager tags to be bound to the instance. Tag keys and values have the same definition as resource manager tags. Keys must be in the format tagKeys/{tag_key_id}, and values are in the format tagValues/456. The field is ignored (both PUT & PATCH) when empty.

resource_policies: Vec<String>

Resource policies (names, not URLs) applied to instances created from these properties. Note that for MachineImage, this is not supported yet.

scheduling: Option<Scheduling>

Specifies the scheduling options for the instances that are created from these properties.

service_accounts: Vec<ServiceAccount>

A list of service accounts with specified scopes. Access tokens for these service accounts are available to the instances that are created from these properties. Use metadata queries to obtain the access tokens for these instances.

shielded_instance_config: Option<ShieldedInstanceConfig>

Note that for MachineImage, this is not supported yet.

tags: Option<Tags>

A list of tags to apply to the instances that are created from these properties. The tags identify valid sources or targets for network firewalls. The setTags method can modify this list of tags. Each tag within the list must comply with RFC1035.

Implementations

impl InstanceProperties

pub fn private_ipv6_google_access(&self) -> &str

Returns the value of private_ipv6_google_access, or the default value if private_ipv6_google_access is unset.

pub fn machine_type(&self) -> &str

Returns the value of machine_type, or the default value if machine_type is unset.

pub fn key_revocation_action_type(&self) -> &str

Returns the value of key_revocation_action_type, or the default value if key_revocation_action_type is unset.

pub fn min_cpu_platform(&self) -> &str

Returns the value of min_cpu_platform, or the default value if min_cpu_platform is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn can_ip_forward(&self) -> bool

Returns the value of can_ip_forward, or the default value if can_ip_forward is unset.