Struct google_api_proto::google::cloud::compute::v1::Network

pub struct Network {

    pub i_pv4_range: Option<String>,
    pub auto_create_subnetworks: Option<bool>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub enable_ula_internal_ipv6: Option<bool>,
    pub firewall_policy: Option<String>,
    pub gateway_i_pv4: Option<String>,
    pub id: Option<u64>,
    pub internal_ipv6_range: Option<String>,
    pub kind: Option<String>,
    pub mtu: Option<i32>,
    pub name: Option<String>,
    pub network_firewall_policy_enforcement_order: Option<String>,
    pub peerings: Vec<NetworkPeering>,
    pub routing_config: Option<NetworkRoutingConfig>,
    pub self_link: Option<String>,
    pub self_link_with_id: Option<String>,
    pub subnetworks: Vec<String>,

}

Represents a VPC Network resource. Networks connect resources to each other and to the internet. For more information, read Virtual Private Cloud (VPC) Network.

Fields
i_pv4_range: Option<String>

Deprecated in favor of subnet mode networks. The range of internal addresses that are legal on this network. This range is a CIDR specification, for example: 192.168.0.0/16. Provided by the client when the network is created.

auto_create_subnetworks: Option<bool>

Must be set to create a VPC network. If not set, a legacy network is created. When set to true, the VPC network is created in auto mode. When set to false, the VPC network is created in custom mode. An auto mode VPC network starts with one subnet per region. Each subnet has a predetermined range as described in Auto mode VPC network IP ranges. For custom mode VPC networks, you can add subnets using the subnetworks insert method.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this field when you create the resource.

enable_ula_internal_ipv6: Option<bool>

Enable ULA internal ipv6 on this network. Enabling this feature will assign a /48 from google defined ULA prefix fd20::/20. .

firewall_policy: Option<String>

[Output Only] URL of the firewall policy the network is associated with.

gateway_i_pv4: Option<String>

[Output Only] The gateway address for default routing out of the network, selected by Google Cloud.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

internal_ipv6_range: Option<String>

When enabling ula internal ipv6, caller optionally can specify the /48 range they want from the google defined ULA prefix fd20::/20. The input must be a valid /48 ULA IPv6 address and must be within the fd20::/20. Operation will fail if the speficied /48 is already in used by another resource. If the field is not speficied, then a /48 range will be randomly allocated from fd20::/20 and returned via this field. .

kind: Option<String>

[Output Only] Type of the resource. Always compute#network for networks.

mtu: Option<i32>

Maximum Transmission Unit in bytes. The minimum value for this field is 1300 and the maximum value is 8896. The suggested value is 1500, which is the default MTU used on the Internet, or 8896 if you want to use Jumbo frames. If unspecified, the value defaults to 1460.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])?. The first character must be a lowercase letter, and all following characters (except for the last character) must be a dash, lowercase letter, or digit. The last character must be a lowercase letter or digit.

network_firewall_policy_enforcement_order: Option<String>

The network firewall policy enforcement order. Can be either AFTER_CLASSIC_FIREWALL or BEFORE_CLASSIC_FIREWALL. Defaults to AFTER_CLASSIC_FIREWALL if the field is not specified. Check the NetworkFirewallPolicyEnforcementOrder enum for the list of possible values.

peerings: Vec<NetworkPeering>

[Output Only] A list of network peerings for the resource.

routing_config: Option<NetworkRoutingConfig>

The network-level routing configuration for this network. Used by Cloud Router to determine what type of network-wide routing behavior to enforce.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

self_link_with_id: Option<String>

[Output Only] Server-defined URL for this resource with the resource id.

subnetworks: Vec<String>

[Output Only] Server-defined fully-qualified URLs for all subnetworks in this VPC network.

Implementations

impl Network

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn mtu(&self) -> i32

Returns the value of mtu, or the default value if mtu is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn network_firewall_policy_enforcement_order(&self) -> &str

Returns the value of network_firewall_policy_enforcement_order, or the default value if network_firewall_policy_enforcement_order is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn self_link_with_id(&self) -> &str

Returns the value of self_link_with_id, or the default value if self_link_with_id is unset.

pub fn i_pv4_range(&self) -> &str

Returns the value of i_pv4_range, or the default value if i_pv4_range is unset.

pub fn gateway_i_pv4(&self) -> &str

Returns the value of gateway_i_pv4, or the default value if gateway_i_pv4 is unset.

pub fn auto_create_subnetworks(&self) -> bool

Returns the value of auto_create_subnetworks, or the default value if auto_create_subnetworks is unset.

pub fn internal_ipv6_range(&self) -> &str

Returns the value of internal_ipv6_range, or the default value if internal_ipv6_range is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn enable_ula_internal_ipv6(&self) -> bool

Returns the value of enable_ula_internal_ipv6, or the default value if enable_ula_internal_ipv6 is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn firewall_policy(&self) -> &str

Returns the value of firewall_policy, or the default value if firewall_policy is unset.