Struct google_api_proto::google::cloud::compute::v1::NetworkEndpointGroup

pub struct NetworkEndpointGroup {

    pub annotations: BTreeMap<String, String>,
    pub app_engine: Option<NetworkEndpointGroupAppEngine>,
    pub cloud_function: Option<NetworkEndpointGroupCloudFunction>,
    pub cloud_run: Option<NetworkEndpointGroupCloudRun>,
    pub creation_timestamp: Option<String>,
    pub default_port: Option<i32>,
    pub description: Option<String>,
    pub id: Option<u64>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub network_endpoint_type: Option<String>,
    pub psc_data: Option<NetworkEndpointGroupPscData>,
    pub psc_target_service: Option<String>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub size: Option<i32>,
    pub subnetwork: Option<String>,
    pub zone: Option<String>,

}

Represents a collection of network endpoints. A network endpoint group (NEG) defines how a set of endpoints should be reached, whether they are reachable, and where they are located. For more information about using NEGs for different use cases, see Network endpoint groups overview.

Fields
annotations: BTreeMap<String, String>

Metadata defined as annotations on the network endpoint group.

app_engine: Option<NetworkEndpointGroupAppEngine>

Only valid when networkEndpointType is SERVERLESS. Only one of cloudRun, appEngine or cloudFunction may be set.

cloud_function: Option<NetworkEndpointGroupCloudFunction>

Only valid when networkEndpointType is SERVERLESS. Only one of cloudRun, appEngine or cloudFunction may be set.

cloud_run: Option<NetworkEndpointGroupCloudRun>

Only valid when networkEndpointType is SERVERLESS. Only one of cloudRun, appEngine or cloudFunction may be set.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

default_port: Option<i32>

The default port used if the port number is not specified in the network endpoint. If the network endpoint type is either GCE_VM_IP, SERVERLESS or PRIVATE_SERVICE_CONNECT, this field must not be specified.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#networkEndpointGroup for network endpoint group.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network: Option<String>

The URL of the network to which all network endpoints in the NEG belong. Uses default project network if unspecified.

network_endpoint_type: Option<String>

Type of network endpoints in this network endpoint group. Can be one of GCE_VM_IP, GCE_VM_IP_PORT, NON_GCP_PRIVATE_IP_PORT, INTERNET_FQDN_PORT, INTERNET_IP_PORT, SERVERLESS, PRIVATE_SERVICE_CONNECT, GCE_VM_IP_PORTMAP. Check the NetworkEndpointType enum for the list of possible values.

psc_data: Option<NetworkEndpointGroupPscData>psc_target_service: Option<String>

The target service url used to set up private service connection to a Google API or a PSC Producer Service Attachment. An example value is: asia-northeast3-cloudkms.googleapis.com

region: Option<String>

[Output Only] The URL of the region where the network endpoint group is located.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

size: Option<i32>

[Output only] Number of network endpoints in the network endpoint group.

subnetwork: Option<String>

Optional URL of the subnetwork to which all network endpoints in the NEG belong.

zone: Option<String>

[Output Only] The URL of the zone where the network endpoint group is located.

Implementations

impl NetworkEndpointGroup

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn size(&self) -> i32

Returns the value of size, or the default value if size is unset.

pub fn zone(&self) -> &str

Returns the value of zone, or the default value if zone is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn network_endpoint_type(&self) -> &str

Returns the value of network_endpoint_type, or the default value if network_endpoint_type is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn network(&self) -> &str

Returns the value of network, or the default value if network is unset.

pub fn psc_target_service(&self) -> &str

Returns the value of psc_target_service, or the default value if psc_target_service is unset.

pub fn subnetwork(&self) -> &str

Returns the value of subnetwork, or the default value if subnetwork is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn default_port(&self) -> i32

Returns the value of default_port, or the default value if default_port is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.