Struct google_api_proto::google::cloud::compute::v1::OutlierDetection

pub struct OutlierDetection {
    pub base_ejection_time: Option<Duration>,
    pub consecutive_errors: Option<i32>,
    pub consecutive_gateway_failure: Option<i32>,
    pub enforcing_consecutive_errors: Option<i32>,
    pub enforcing_consecutive_gateway_failure: Option<i32>,
    pub enforcing_success_rate: Option<i32>,
    pub interval: Option<Duration>,
    pub max_ejection_percent: Option<i32>,
    pub success_rate_minimum_hosts: Option<i32>,
    pub success_rate_request_volume: Option<i32>,
    pub success_rate_stdev_factor: Option<i32>,
}

Settings controlling the eviction of unhealthy hosts from the load balancing pool for the backend service.

Fields
base_ejection_time: Option<Duration>

The base time that a backend endpoint is ejected for. Defaults to 30000ms or 30s. After a backend endpoint is returned back to the load balancing pool, it can be ejected again in another ejection analysis. Thus, the total ejection time is equal to the base ejection time multiplied by the number of times the backend endpoint has been ejected. Defaults to 30000ms or 30s.

consecutive_errors: Option<i32>

Number of consecutive errors before a backend endpoint is ejected from the load balancing pool. When the backend endpoint is accessed over HTTP, a 5xx return code qualifies as an error. Defaults to 5.

consecutive_gateway_failure: Option<i32>

The number of consecutive gateway failures (502, 503, 504 status or connection errors that are mapped to one of those status codes) before a consecutive gateway failure ejection occurs. Defaults to 3.

enforcing_consecutive_errors: Option<i32>

The percentage chance that a backend endpoint will be ejected when an outlier status is detected through consecutive 5xx. This setting can be used to disable ejection or to ramp it up slowly. Defaults to 0.

enforcing_consecutive_gateway_failure: Option<i32>

The percentage chance that a backend endpoint will be ejected when an outlier status is detected through consecutive gateway failures. This setting can be used to disable ejection or to ramp it up slowly. Defaults to 100.

enforcing_success_rate: Option<i32>

The percentage chance that a backend endpoint will be ejected when an outlier status is detected through success rate statistics. This setting can be used to disable ejection or to ramp it up slowly. Defaults to 100. Not supported when the backend service uses Serverless NEG.

interval: Option<Duration>

Time interval between ejection analysis sweeps. This can result in both new ejections and backend endpoints being returned to service. The interval is equal to the number of seconds as defined in outlierDetection.interval.seconds plus the number of nanoseconds as defined in outlierDetection.interval.nanos. Defaults to 1 second.

max_ejection_percent: Option<i32>

Maximum percentage of backend endpoints in the load balancing pool for the backend service that can be ejected if the ejection conditions are met. Defaults to 50%.

success_rate_minimum_hosts: Option<i32>

The number of backend endpoints in the load balancing pool that must have enough request volume to detect success rate outliers. If the number of backend endpoints is fewer than this setting, outlier detection via success rate statistics is not performed for any backend endpoint in the load balancing pool. Defaults to 5. Not supported when the backend service uses Serverless NEG.

success_rate_request_volume: Option<i32>

The minimum number of total requests that must be collected in one interval (as defined by the interval duration above) to include this backend endpoint in success rate based outlier detection. If the volume is lower than this setting, outlier detection via success rate statistics is not performed for that backend endpoint. Defaults to 100. Not supported when the backend service uses Serverless NEG.

success_rate_stdev_factor: Option<i32>

This factor is used to determine the ejection threshold for success rate outlier ejection. The ejection threshold is the difference between the mean success rate, and the product of this factor and the standard deviation of the mean success rate: mean - (stdev * successRateStdevFactor). This factor is divided by a thousand to get a double. That is, if the desired factor is 1.9, the runtime value should be 1900. Defaults to 1900. Not supported when the backend service uses Serverless NEG.

Implementations

impl OutlierDetection

pub fn max_ejection_percent(&self) -> i32

Returns the value of max_ejection_percent, or the default value if max_ejection_percent is unset.

pub fn success_rate_stdev_factor(&self) -> i32

Returns the value of success_rate_stdev_factor, or the default value if success_rate_stdev_factor is unset.

pub fn enforcing_success_rate(&self) -> i32

Returns the value of enforcing_success_rate, or the default value if enforcing_success_rate is unset.

pub fn enforcing_consecutive_errors(&self) -> i32

Returns the value of enforcing_consecutive_errors, or the default value if enforcing_consecutive_errors is unset.

pub fn success_rate_request_volume(&self) -> i32

Returns the value of success_rate_request_volume, or the default value if success_rate_request_volume is unset.

pub fn consecutive_errors(&self) -> i32

Returns the value of consecutive_errors, or the default value if consecutive_errors is unset.

pub fn enforcing_consecutive_gateway_failure(&self) -> i32

Returns the value of enforcing_consecutive_gateway_failure, or the default value if enforcing_consecutive_gateway_failure is unset.

pub fn consecutive_gateway_failure(&self) -> i32

Returns the value of consecutive_gateway_failure, or the default value if consecutive_gateway_failure is unset.

pub fn success_rate_minimum_hosts(&self) -> i32

Returns the value of success_rate_minimum_hosts, or the default value if success_rate_minimum_hosts is unset.