Struct google_api_proto::google::cloud::compute::v1::PublicDelegatedPrefix

pub struct PublicDelegatedPrefix {

    pub allocatable_prefix_length: Option<i32>,
    pub byoip_api_version: Option<String>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub fingerprint: Option<String>,
    pub id: Option<u64>,
    pub ip_cidr_range: Option<String>,
    pub is_live_migration: Option<bool>,
    pub kind: Option<String>,
    pub mode: Option<String>,
    pub name: Option<String>,
    pub parent_prefix: Option<String>,
    pub public_delegated_sub_prefixs: Vec<PublicDelegatedPrefixPublicDelegatedSubPrefix>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub status: Option<String>,

}

A PublicDelegatedPrefix resource represents an IP block within a PublicAdvertisedPrefix that is configured within a single cloud scope (global or region). IPs in the block can be allocated to resources within that scope. Public delegated prefixes may be further broken up into smaller IP blocks in the same scope as the parent block.

Fields
allocatable_prefix_length: Option<i32>

The allocatable prefix length supported by this public delegated prefix. This field is optional and cannot be set for prefixes in DELEGATION mode. It cannot be set for IPv4 prefixes either, and it always defaults to 32.

byoip_api_version: Option<String>

[Output Only] The version of BYOIP API. Check the ByoipApiVersion enum for the list of possible values.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a new PublicDelegatedPrefix. An up-to-date fingerprint must be provided in order to update the PublicDelegatedPrefix, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a PublicDelegatedPrefix.

id: Option<u64>

[Output Only] The unique identifier for the resource type. The server generates this identifier.

ip_cidr_range: Option<String>

The IP address range, in CIDR format, represented by this public delegated prefix.

is_live_migration: Option<bool>

If true, the prefix will be live migrated.

kind: Option<String>

[Output Only] Type of the resource. Always compute#publicDelegatedPrefix for public delegated prefixes.

mode: Option<String>

The public delegated prefix mode for IPv6 only. Check the Mode enum for the list of possible values.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

parent_prefix: Option<String>

The URL of parent prefix. Either PublicAdvertisedPrefix or PublicDelegatedPrefix.

public_delegated_sub_prefixs: Vec<PublicDelegatedPrefixPublicDelegatedSubPrefix>

The list of sub public delegated prefixes that exist for this public delegated prefix.

region: Option<String>

[Output Only] URL of the region where the public delegated prefix resides. This field applies only to the region resource. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

status: Option<String>

[Output Only] The status of the public delegated prefix, which can be one of following values: - INITIALIZING The public delegated prefix is being initialized and addresses cannot be created yet. - READY_TO_ANNOUNCE The public delegated prefix is a live migration prefix and is active. - ANNOUNCED The public delegated prefix is active. - DELETING The public delegated prefix is being deprovsioned. Check the Status enum for the list of possible values.

Implementations

impl PublicDelegatedPrefix

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn mode(&self) -> &str

Returns the value of mode, or the default value if mode is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn parent_prefix(&self) -> &str

Returns the value of parent_prefix, or the default value if parent_prefix is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn allocatable_prefix_length(&self) -> i32

Returns the value of allocatable_prefix_length, or the default value if allocatable_prefix_length is unset.

pub fn ip_cidr_range(&self) -> &str

Returns the value of ip_cidr_range, or the default value if ip_cidr_range is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn byoip_api_version(&self) -> &str

Returns the value of byoip_api_version, or the default value if byoip_api_version is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn is_live_migration(&self) -> bool

Returns the value of is_live_migration, or the default value if is_live_migration is unset.