Struct google_api_proto::google::cloud::compute::v1::Scheduling

pub struct Scheduling {
    pub automatic_restart: Option<bool>,
    pub instance_termination_action: Option<String>,
    pub local_ssd_recovery_timeout: Option<Duration>,
    pub location_hint: Option<String>,
    pub min_node_cpus: Option<i32>,
    pub node_affinities: Vec<SchedulingNodeAffinity>,
    pub on_host_maintenance: Option<String>,
    pub preemptible: Option<bool>,
    pub provisioning_model: Option<String>,
}

Sets the scheduling options for an Instance.

Fields
automatic_restart: Option<bool>

Specifies whether the instance should be automatically restarted if it is terminated by Compute Engine (not terminated by a user). You can only set the automatic restart option for standard instances. Preemptible instances cannot be automatically restarted. By default, this is set to true so an instance is automatically restarted if it is terminated by Compute Engine.

instance_termination_action: Option<String>

Specifies the termination action for the instance. Check the InstanceTerminationAction enum for the list of possible values.

local_ssd_recovery_timeout: Option<Duration>

Specifies the maximum amount of time a Local Ssd Vm should wait while recovery of the Local Ssd state is attempted. Its value should be in between 0 and 168 hours with hour granularity and the default value being 1 hour.

location_hint: Option<String>

An opaque location hint used to place the instance close to other resources. This field is for use by internal tools that use the public API.

min_node_cpus: Option<i32>

The minimum number of virtual CPUs this instance will consume when running on a sole-tenant node.

node_affinities: Vec<SchedulingNodeAffinity>

A set of node affinity and anti-affinity configurations. Refer to Configuring node affinity for more information. Overrides reservationAffinity.

on_host_maintenance: Option<String>

Defines the maintenance behavior for this instance. For standard instances, the default behavior is MIGRATE. For preemptible instances, the default and only possible behavior is TERMINATE. For more information, see Set VM host maintenance policy. Check the OnHostMaintenance enum for the list of possible values.

preemptible: Option<bool>

Defines whether the instance is preemptible. This can only be set during instance creation or while the instance is stopped and therefore, in a TERMINATED state. See Instance Life Cycle for more information on the possible instance states.

provisioning_model: Option<String>

Specifies the provisioning model of the instance. Check the ProvisioningModel enum for the list of possible values.

Implementations

impl Scheduling

pub fn provisioning_model(&self) -> &str

Returns the value of provisioning_model, or the default value if provisioning_model is unset.

pub fn on_host_maintenance(&self) -> &str

Returns the value of on_host_maintenance, or the default value if on_host_maintenance is unset.

pub fn instance_termination_action(&self) -> &str

Returns the value of instance_termination_action, or the default value if instance_termination_action is unset.

pub fn min_node_cpus(&self) -> i32

Returns the value of min_node_cpus, or the default value if min_node_cpus is unset.

pub fn preemptible(&self) -> bool

Returns the value of preemptible, or the default value if preemptible is unset.

pub fn location_hint(&self) -> &str

Returns the value of location_hint, or the default value if location_hint is unset.

pub fn automatic_restart(&self) -> bool

Returns the value of automatic_restart, or the default value if automatic_restart is unset.