Struct google_api_proto::google::cloud::compute::v1::VpnTunnel

pub struct VpnTunnel {

    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub detailed_status: Option<String>,
    pub id: Option<u64>,
    pub ike_version: Option<i32>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub local_traffic_selector: Vec<String>,
    pub name: Option<String>,
    pub peer_external_gateway: Option<String>,
    pub peer_external_gateway_interface: Option<i32>,
    pub peer_gcp_gateway: Option<String>,
    pub peer_ip: Option<String>,
    pub region: Option<String>,
    pub remote_traffic_selector: Vec<String>,
    pub router: Option<String>,
    pub self_link: Option<String>,
    pub shared_secret: Option<String>,
    pub shared_secret_hash: Option<String>,
    pub status: Option<String>,
    pub target_vpn_gateway: Option<String>,
    pub vpn_gateway: Option<String>,
    pub vpn_gateway_interface: Option<i32>,

}

Represents a Cloud VPN Tunnel resource. For more information about VPN, read the the Cloud VPN Overview.

Fields
creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

detailed_status: Option<String>

[Output Only] Detailed status message for the VPN tunnel.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

ike_version: Option<i32>

IKE protocol version to use when establishing the VPN tunnel with the peer VPN gateway. Acceptable IKE versions are 1 or 2. The default version is 2.

kind: Option<String>

[Output Only] Type of resource. Always compute#vpnTunnel for VPN tunnels.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this VpnTunnel, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a VpnTunnel.

labels: BTreeMap<String, String>

Labels for this resource. These can only be added or modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.

local_traffic_selector: Vec<String>

Local traffic selector to use when establishing the VPN tunnel with the peer VPN gateway. The value should be a CIDR formatted string, for example: 192.168.0.0/16. The ranges must be disjoint. Only IPv4 is supported.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

peer_external_gateway: Option<String>

URL of the peer side external VPN gateway to which this VPN tunnel is connected. Provided by the client when the VPN tunnel is created. This field is exclusive with the field peerGcpGateway.

peer_external_gateway_interface: Option<i32>

The interface ID of the external VPN gateway to which this VPN tunnel is connected. Provided by the client when the VPN tunnel is created. Possible values are: 0, 1, 2, 3. The number of IDs in use depends on the external VPN gateway redundancy type.

peer_gcp_gateway: Option<String>

URL of the peer side HA VPN gateway to which this VPN tunnel is connected. Provided by the client when the VPN tunnel is created. This field can be used when creating highly available VPN from VPC network to VPC network, the field is exclusive with the field peerExternalGateway. If provided, the VPN tunnel will automatically use the same vpnGatewayInterface ID in the peer Google Cloud VPN gateway.

peer_ip: Option<String>

IP address of the peer VPN gateway. Only IPv4 is supported.

region: Option<String>

[Output Only] URL of the region where the VPN tunnel resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

remote_traffic_selector: Vec<String>

Remote traffic selectors to use when establishing the VPN tunnel with the peer VPN gateway. The value should be a CIDR formatted string, for example: 192.168.0.0/16. The ranges should be disjoint. Only IPv4 is supported.

router: Option<String>

URL of the router resource to be used for dynamic routing.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

shared_secret: Option<String>

Shared secret used to set the secure session between the Cloud VPN gateway and the peer VPN gateway.

shared_secret_hash: Option<String>

Hash of the shared secret.

status: Option<String>

[Output Only] The status of the VPN tunnel, which can be one of the following: - PROVISIONING: Resource is being allocated for the VPN tunnel. - WAITING_FOR_FULL_CONFIG: Waiting to receive all VPN-related configs from the user. Network, TargetVpnGateway, VpnTunnel, ForwardingRule, and Route resources are needed to setup the VPN tunnel. - FIRST_HANDSHAKE: Successful first handshake with the peer VPN. - ESTABLISHED: Secure session is successfully established with the peer VPN. - NETWORK_ERROR: Deprecated, replaced by NO_INCOMING_PACKETS - AUTHORIZATION_ERROR: Auth error (for example, bad shared secret). - NEGOTIATION_FAILURE: Handshake failed. - DEPROVISIONING: Resources are being deallocated for the VPN tunnel. - FAILED: Tunnel creation has failed and the tunnel is not ready to be used. - NO_INCOMING_PACKETS: No incoming packets from peer. - REJECTED: Tunnel configuration was rejected, can be result of being denied access. - ALLOCATING_RESOURCES: Cloud VPN is in the process of allocating all required resources. - STOPPED: Tunnel is stopped due to its Forwarding Rules being deleted for Classic VPN tunnels or the project is in frozen state. - PEER_IDENTITY_MISMATCH: Peer identity does not match peer IP, probably behind NAT. - TS_NARROWING_NOT_ALLOWED: Traffic selector narrowing not allowed for an HA-VPN tunnel. Check the Status enum for the list of possible values.

target_vpn_gateway: Option<String>

URL of the Target VPN gateway with which this VPN tunnel is associated. Provided by the client when the VPN tunnel is created.

vpn_gateway: Option<String>

URL of the VPN gateway with which this VPN tunnel is associated. Provided by the client when the VPN tunnel is created. This must be used (instead of target_vpn_gateway) if a High Availability VPN gateway resource is created.

vpn_gateway_interface: Option<i32>

The interface ID of the VPN gateway with which this VPN tunnel is associated. Possible values are: 0, 1.

Implementations

impl VpnTunnel

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn vpn_gateway_interface(&self) -> i32

Returns the value of vpn_gateway_interface, or the default value if vpn_gateway_interface is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn router(&self) -> &str

Returns the value of router, or the default value if router is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn ike_version(&self) -> i32

Returns the value of ike_version, or the default value if ike_version is unset.

pub fn peer_gcp_gateway(&self) -> &str

Returns the value of peer_gcp_gateway, or the default value if peer_gcp_gateway is unset.

pub fn detailed_status(&self) -> &str

Returns the value of detailed_status, or the default value if detailed_status is unset.

pub fn shared_secret(&self) -> &str

Returns the value of shared_secret, or the default value if shared_secret is unset.

pub fn peer_ip(&self) -> &str

Returns the value of peer_ip, or the default value if peer_ip is unset.

pub fn peer_external_gateway(&self) -> &str

Returns the value of peer_external_gateway, or the default value if peer_external_gateway is unset.

pub fn shared_secret_hash(&self) -> &str

Returns the value of shared_secret_hash, or the default value if shared_secret_hash is unset.

pub fn vpn_gateway(&self) -> &str

Returns the value of vpn_gateway, or the default value if vpn_gateway is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn peer_external_gateway_interface(&self) -> i32

Returns the value of peer_external_gateway_interface, or the default value if peer_external_gateway_interface is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn target_vpn_gateway(&self) -> &str

Returns the value of target_vpn_gateway, or the default value if target_vpn_gateway is unset.