Module google_api_proto::google::cloud::compute::v1::update_instance_request

Nested message and enum types in UpdateInstanceRequest.

Enums

MinimalAction

Specifies the action to take when updating an instance even if the updated properties do not require it. If not specified, then Compute Engine acts based on the minimum action that the updated properties require. Additional supported values which may be not listed in the enum directly due to technical reasons: INVALID NO_EFFECT REFRESH RESTART

MostDisruptiveAllowedAction

Specifies the most disruptive action that can be taken on the instance as part of the update. Compute Engine returns an error if the instance properties require a more disruptive action as part of the instance update. Valid options from lowest to highest are NO_EFFECT, REFRESH, and RESTART. Additional supported values which may be not listed in the enum directly due to technical reasons: INVALID NO_EFFECT REFRESH RESTART