Module google_api_proto::google::cloud::confidentialcomputing::v1alpha1

Modules

confidential_computing_client

Generated client implementations.

tpm_attestation

Nested message and enum types in TpmAttestation.

Structs

Challenge

A Challenge from the server used to guarantee freshness of attestations

CreateChallengeRequest

Message for creating a Challenge

GcpCredentials

Credentials issued by GCP which are linked to the platform attestation. These will be verified server-side as part of attestaion verification.

TpmAttestation

TPM2 data containing everything necessary to validate any platform state measured into the TPM.

VerifyAttestationRequest

A request for an OIDC token, providing all the necessary information needed for this service to verify the plaform state of the requestor.

VerifyAttestationResponse

A response once an attestation has been successfully verified, containing a signed OIDC token.