Module google_api_proto::google::cloud::connectors::v1

Modules

auth_config

Nested message and enum types in AuthConfig.

config_variable

Nested message and enum types in ConfigVariable.

config_variable_template

Nested message and enum types in ConfigVariableTemplate.

connection_schema_metadata

Nested message and enum types in ConnectionSchemaMetadata.

connection_status

Nested message and enum types in ConnectionStatus.

connectors_client

Generated client implementations.

destination

Nested message and enum types in Destination.

egress_control_config

Nested message and enum types in EgressControlConfig.

extraction_rule

Nested message and enum types in ExtractionRule.

role_grant

Nested message and enum types in RoleGrant.

runtime_action_schema

Nested message and enum types in RuntimeActionSchema.

runtime_config

Nested message and enum types in RuntimeConfig.

runtime_entity_schema

Nested message and enum types in RuntimeEntitySchema.

ssl_config

Nested message and enum types in SslConfig.

Structs

AuthConfig

AuthConfig defines details of a authentication type.

AuthConfigTemplate

AuthConfigTemplate defines required field over an authentication type.

AuthorizationCodeLink

This configuration captures the details required to render an authorization link for the OAuth Authorization Code Flow.

ConfigVariable

ConfigVariable represents a configuration variable present in a Connection. or AuthConfig.

ConfigVariableTemplate

ConfigVariableTemplate provides metadata about a ConfigVariable that is used in a Connection.

Connection

Connection represents an instance of connector.

ConnectionSchemaMetadata

ConnectionSchemaMetadata is the singleton resource of each connection. It includes the entity and action names of runtime resources exposed by a connection backend.

ConnectionStatus

ConnectionStatus indicates the state of the connection.

Connector

Connectors indicates a specific connector type, e.x. Salesforce, SAP etc.

ConnectorVersion

ConnectorVersion indicates a specific version of a connector.

CreateConnectionRequest

Request message for ConnectorsService.CreateConnection

DeleteConnectionRequest

Request message for ConnectorsService.DeleteConnection.

Destination

DestinationConfig

Define the Connectors target endpoint.

EgressControlConfig

Egress control config for connector runtime. These configurations define the rules to identify which outbound domains/hosts needs to be whitelisted. It may be a static information for a particular connector version or it is derived from the configurations provided by the customer in Connection resource.

EnumOption

EnumOption definition

ExtractionRule

Extraction Rule.

ExtractionRules

Extraction Rules to identity the backends from customer provided configuration in Connection resource.

GetConnectionRequest

Request message for ConnectorsService.GetConnection

GetConnectionSchemaMetadataRequest

Request message for ConnectorsService.GetConnectionSchemaMetadata.

GetConnectorRequest

Request message for Connectors.GetConnector.

GetConnectorVersionRequest

Request message for Connectors.GetConnectorVersion.

GetGlobalSettingsRequest

Request for [GetGlobalSettingsRequest].

GetProviderRequest

Request message for Connectors.GetProvider.

GetRuntimeConfigRequest

Request message for Connectors.GetRuntimeConfig.

ListConnectionsRequest

Request message for ConnectorsService.ListConnections

ListConnectionsResponse

Response message for ConnectorsService.ListConnections

ListConnectorVersionsRequest

Request message for Connectors.ListConnectorVersions.

ListConnectorVersionsResponse

Response message for Connectors.ListConnectorVersions.

ListConnectorsRequest

Request message for Connectors.ListConnectors.

ListConnectorsResponse

Response message for Connectors.ListConnectors.

ListProvidersRequest

Request message for Connectors.ListProviders.

ListProvidersResponse

Response message for Connectors.ListProviders.

ListRuntimeActionSchemasRequest

Request message for ConnectorsService.ListRuntimeActionSchemas. For filter, only action field is supported with literal equality operator. Accepted filter example: action=“approveOrder” Wildcards are not supported in the filter currently.

ListRuntimeActionSchemasResponse

Response message for ConnectorsService.ListRuntimeActionSchemas.

ListRuntimeEntitySchemasRequest

Request message for ConnectorsService.ListRuntimeEntitySchemas. For filter, only entity field is supported with literal equality operator. Accepted filter example: entity=“Order” Wildcards are not supported in the filter currently.

ListRuntimeEntitySchemasResponse

Response message for ConnectorsService.ListRuntimeEntitySchemas.

LockConfig

Determines whether or no a connection is locked. If locked, a reason must be specified.

NodeConfig

Node configuration for the connection.

OperationMetadata

Represents the metadata of the long-running operation.

Provider

Provider indicates the owner who provides the connectors.

RefreshConnectionSchemaMetadataRequest

Request message for ConnectorsService.RefreshConnectionSchemaMetadata.

RoleGrant

This configuration defines all the Cloud IAM roles that needs to be granted to a particular GCP resource for the selected prinicpal like service account. These configurations will let UI display to customers what IAM roles need to be granted by them. Or these configurations can be used by the UI to render a ‘grant’ button to do the same on behalf of the user.

RuntimeActionSchema

Schema of a runtime action.

RuntimeConfig

RuntimeConfig is the singleton resource of each location. It includes generic resource configs consumed by control plane and runtime plane like: pub/sub topic/subscription resource name, Cloud Storage location storing schema etc.

RuntimeEntitySchema

Schema of a runtime entity.

Secret

Secret provides a reference to entries in Secret Manager.

Settings

Global Settings details.

SslConfig

SSL Configuration of a connection

SslConfigTemplate

Ssl config details of a connector version

SupportedRuntimeFeatures

Supported runtime features of a connector version. This is passed to the management layer to add a new connector version by the connector developer. Details about how this proto is passed to the management layer is covered in this doc - go/runtime-manifest.

UpdateConnectionRequest

Request message for ConnectorsService.UpdateConnection

Enums

AuthType

AuthType defines different authentication types.

CertType

Enum for Cert Types

ConnectionView

Enum to control which fields should be included in the response.

ConnectorVersionView

Enum to control which fields should be included in the response.

DataType

All possible data types of a entity or action field.

LaunchStage

LaunchStage is a enum to indicate launch stage: PREVIEW, GA, DEPRECATED, PRIVATE_PREVIEW.

SslType

Enum for controlling the SSL Type (TLS/MTLS)