Struct google_api_proto::google::cloud::connectors::v1::connectors_client::ConnectorsClient

pub struct ConnectorsClient<T> { /* private fields */ }

Connectors is the interface for managing Connectors.

Implementations

impl<T> ConnectorsClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> ConnectorsClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_connections( &mut self, request: impl IntoRequest<ListConnectionsRequest>, ) -> Result<Response<ListConnectionsResponse>, Status>

Lists Connections in a given project and location.

pub async fn get_connection( &mut self, request: impl IntoRequest<GetConnectionRequest>, ) -> Result<Response<Connection>, Status>

Gets details of a single Connection.

pub async fn create_connection( &mut self, request: impl IntoRequest<CreateConnectionRequest>, ) -> Result<Response<Operation>, Status>

Creates a new Connection in a given project and location.

pub async fn update_connection( &mut self, request: impl IntoRequest<UpdateConnectionRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Connection.

pub async fn delete_connection( &mut self, request: impl IntoRequest<DeleteConnectionRequest>, ) -> Result<Response<Operation>, Status>

Deletes a single Connection.

pub async fn list_providers( &mut self, request: impl IntoRequest<ListProvidersRequest>, ) -> Result<Response<ListProvidersResponse>, Status>

Lists Providers in a given project and location.

pub async fn get_provider( &mut self, request: impl IntoRequest<GetProviderRequest>, ) -> Result<Response<Provider>, Status>

Gets details of a provider.

pub async fn list_connectors( &mut self, request: impl IntoRequest<ListConnectorsRequest>, ) -> Result<Response<ListConnectorsResponse>, Status>

Lists Connectors in a given project and location.

pub async fn get_connector( &mut self, request: impl IntoRequest<GetConnectorRequest>, ) -> Result<Response<Connector>, Status>

Gets details of a single Connector.

pub async fn list_connector_versions( &mut self, request: impl IntoRequest<ListConnectorVersionsRequest>, ) -> Result<Response<ListConnectorVersionsResponse>, Status>

Lists Connector Versions in a given project and location.

pub async fn get_connector_version( &mut self, request: impl IntoRequest<GetConnectorVersionRequest>, ) -> Result<Response<ConnectorVersion>, Status>

Gets details of a single connector version.

pub async fn get_connection_schema_metadata( &mut self, request: impl IntoRequest<GetConnectionSchemaMetadataRequest>, ) -> Result<Response<ConnectionSchemaMetadata>, Status>

Gets schema metadata of a connection. SchemaMetadata is a singleton resource for each connection.

pub async fn refresh_connection_schema_metadata( &mut self, request: impl IntoRequest<RefreshConnectionSchemaMetadataRequest>, ) -> Result<Response<Operation>, Status>

Refresh runtime schema of a connection.

pub async fn list_runtime_entity_schemas( &mut self, request: impl IntoRequest<ListRuntimeEntitySchemasRequest>, ) -> Result<Response<ListRuntimeEntitySchemasResponse>, Status>

List schema of a runtime entities filtered by entity name.

pub async fn list_runtime_action_schemas( &mut self, request: impl IntoRequest<ListRuntimeActionSchemasRequest>, ) -> Result<Response<ListRuntimeActionSchemasResponse>, Status>

List schema of a runtime actions filtered by action name.

pub async fn get_runtime_config( &mut self, request: impl IntoRequest<GetRuntimeConfigRequest>, ) -> Result<Response<RuntimeConfig>, Status>

Gets the runtimeConfig of a location. RuntimeConfig is a singleton resource for each location.

pub async fn get_global_settings( &mut self, request: impl IntoRequest<GetGlobalSettingsRequest>, ) -> Result<Response<Settings>, Status>

GetGlobalSettings gets settings of a project. GlobalSettings is a singleton resource.