Struct google_api_proto::google::cloud::connectors::v1::ConfigVariableTemplate

pub struct ConfigVariableTemplate {
    pub key: String,
    pub value_type: i32,
    pub display_name: String,
    pub description: String,
    pub validation_regex: String,
    pub required: bool,
    pub role_grant: Option<RoleGrant>,
    pub enum_options: Vec<EnumOption>,
    pub authorization_code_link: Option<AuthorizationCodeLink>,
    pub state: i32,
    pub is_advanced: bool,
}

ConfigVariableTemplate provides metadata about a ConfigVariable that is used in a Connection.

Fields
key: String

Key of the config variable.

value_type: i32

Type of the parameter: string, int, bool etc. consider custom type for the benefit for the validation.

display_name: String

Display name of the parameter.

description: String

Description.

validation_regex: String

Regular expression in RE2 syntax used for validating the value of a ConfigVariable.

required: bool

Flag represents that this ConfigVariable must be provided for a connection.

role_grant: Option<RoleGrant>

Role grant configuration for the config variable.

enum_options: Vec<EnumOption>

Enum options. To be populated if ValueType is ENUM

authorization_code_link: Option<AuthorizationCodeLink>

Authorization code link options. To be populated if ValueType is AUTHORIZATION_CODE

state: i32

State of the config variable.

is_advanced: bool

Indicates if current template is part of advanced settings

Implementations

impl ConfigVariableTemplate

pub fn value_type(&self) -> ValueType

Returns the enum value of value_type, or the default if the field is set to an invalid enum value.

pub fn set_value_type(&mut self, value: ValueType)

Sets value_type to the provided enum value.

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.