Struct google_api_proto::google::cloud::datacatalog::lineage::v1::lineage_client::LineageClient

pub struct LineageClient<T> { /* private fields */ }

Lineage is used to track data flows between assets over time. You can create [LineageEvents][google.cloud.datacatalog.lineage.v1.LineageEvent] to record lineage between multiple sources and a single target, for example, when table data is based on data from multiple tables.

Implementations

impl<T> LineageClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> LineageClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn process_open_lineage_run_event( &mut self, request: impl IntoRequest<ProcessOpenLineageRunEventRequest> ) -> Result<Response<ProcessOpenLineageRunEventResponse>, Status>

Creates new lineage events together with their parents: process and run. Updates the process and run if they already exist. Mapped from Open Lineage specification: https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.json.

pub async fn create_process( &mut self, request: impl IntoRequest<CreateProcessRequest> ) -> Result<Response<Process>, Status>

Creates a new process.

pub async fn update_process( &mut self, request: impl IntoRequest<UpdateProcessRequest> ) -> Result<Response<Process>, Status>

Updates a process.

pub async fn get_process( &mut self, request: impl IntoRequest<GetProcessRequest> ) -> Result<Response<Process>, Status>

Gets the details of the specified process.

pub async fn list_processes( &mut self, request: impl IntoRequest<ListProcessesRequest> ) -> Result<Response<ListProcessesResponse>, Status>

List processes in the given project and location. List order is descending by insertion time.

pub async fn delete_process( &mut self, request: impl IntoRequest<DeleteProcessRequest> ) -> Result<Response<Operation>, Status>

Deletes the process with the specified name.

pub async fn create_run( &mut self, request: impl IntoRequest<CreateRunRequest> ) -> Result<Response<Run>, Status>

Creates a new run.

pub async fn update_run( &mut self, request: impl IntoRequest<UpdateRunRequest> ) -> Result<Response<Run>, Status>

Updates a run.

pub async fn get_run( &mut self, request: impl IntoRequest<GetRunRequest> ) -> Result<Response<Run>, Status>

Gets the details of the specified run.

pub async fn list_runs( &mut self, request: impl IntoRequest<ListRunsRequest> ) -> Result<Response<ListRunsResponse>, Status>

Lists runs in the given project and location. List order is descending by start_time.

pub async fn delete_run( &mut self, request: impl IntoRequest<DeleteRunRequest> ) -> Result<Response<Operation>, Status>

Deletes the run with the specified name.

pub async fn create_lineage_event( &mut self, request: impl IntoRequest<CreateLineageEventRequest> ) -> Result<Response<LineageEvent>, Status>

Creates a new lineage event.

pub async fn get_lineage_event( &mut self, request: impl IntoRequest<GetLineageEventRequest> ) -> Result<Response<LineageEvent>, Status>

Gets details of a specified lineage event.

pub async fn list_lineage_events( &mut self, request: impl IntoRequest<ListLineageEventsRequest> ) -> Result<Response<ListLineageEventsResponse>, Status>

Lists lineage events in the given project and location. The list order is not defined.

pub async fn delete_lineage_event( &mut self, request: impl IntoRequest<DeleteLineageEventRequest> ) -> Result<Response<()>, Status>

Deletes the lineage event with the specified name.

pub async fn search_links( &mut self, request: impl IntoRequest<SearchLinksRequest> ) -> Result<Response<SearchLinksResponse>, Status>

Retrieve a list of links connected to a specific asset. Links represent the data flow between source (upstream) and target (downstream) assets in transformation pipelines. Links are stored in the same project as the Lineage Events that create them.

You can retrieve links in every project where you have the datalineage.events.get permission. The project provided in the URL is used for Billing and Quota.

pub async fn batch_search_link_processes( &mut self, request: impl IntoRequest<BatchSearchLinkProcessesRequest> ) -> Result<Response<BatchSearchLinkProcessesResponse>, Status>

Retrieve information about LineageProcesses associated with specific links. LineageProcesses are transformation pipelines that result in data flowing from source to target assets. Links between assets represent this operation.

If you have specific link names, you can use this method to verify which LineageProcesses contribute to creating those links. See the [SearchLinks][google.cloud.datacatalog.lineage.v1.Lineage.SearchLinks] method for more information on how to retrieve link name.

You can retrieve the LineageProcess information in every project where you have the datalineage.events.get permission. The project provided in the URL is used for Billing and Quota.