Module google_api_proto::google::cloud::datacatalog::v1

Modules

big_query_connection_spec

Nested message and enum types in BigQueryConnectionSpec.

big_query_table_spec

Nested message and enum types in BigQueryTableSpec.

cloud_bigtable_instance_spec

Nested message and enum types in CloudBigtableInstanceSpec.

cloud_sql_big_query_connection_spec

Nested message and enum types in CloudSqlBigQueryConnectionSpec.

column_schema

Nested message and enum types in ColumnSchema.

contacts

Nested message and enum types in Contacts.

data_catalog_client

Generated client implementations.

data_source

Nested message and enum types in DataSource.

database_table_spec

Nested message and enum types in DatabaseTableSpec.

dataset_spec

Nested message and enum types in DatasetSpec.

dump_item

Nested message and enum types in DumpItem.

entry

Nested message and enum types in Entry.

export_taxonomies_request

Nested message and enum types in ExportTaxonomiesRequest.

field_type

Nested message and enum types in FieldType.

import_entries_metadata

Nested message and enum types in ImportEntriesMetadata.

import_entries_request

Nested message and enum types in ImportEntriesRequest.

import_taxonomies_request

Nested message and enum types in ImportTaxonomiesRequest.

lookup_entry_request

Nested message and enum types in LookupEntryRequest.

model_spec

Nested message and enum types in ModelSpec.

physical_schema

Nested message and enum types in PhysicalSchema.

policy_tag_manager_client

Generated client implementations.

policy_tag_manager_serialization_client

Generated client implementations.

reconcile_tags_metadata

Nested message and enum types in ReconcileTagsMetadata.

routine_spec

Nested message and enum types in RoutineSpec.

search_catalog_request

Nested message and enum types in SearchCatalogRequest.

search_catalog_result

Nested message and enum types in SearchCatalogResult.

service_spec

Nested message and enum types in ServiceSpec.

tag

Nested message and enum types in Tag.

tag_field

Nested message and enum types in TagField.

tagged_entry

Nested message and enum types in TaggedEntry.

taxonomy

Nested message and enum types in Taxonomy.

vertex_dataset_spec

Nested message and enum types in VertexDatasetSpec.

vertex_model_source_info

Nested message and enum types in VertexModelSourceInfo.

Structs

BigQueryConnectionSpec

Specification for the BigQuery connection.

BigQueryDateShardedSpec

Specification for a group of BigQuery tables with the \[prefix\]YYYYMMDD name pattern.

BigQueryRoutineSpec

Fields specific for BigQuery routines.

BigQueryTableSpec

Describes a BigQuery table.

BusinessContext

Business Context of the entry.

CloudBigtableInstanceSpec

Specification that applies to Instance entries that are part of CLOUD_BIGTABLE system. (user_specified_type)

CloudBigtableSystemSpec

Specification that applies to all entries that are part of CLOUD_BIGTABLE system (user_specified_type)

CloudSqlBigQueryConnectionSpec

Specification for the BigQuery connection to a Cloud SQL instance.

ColumnSchema

A column within a schema. Columns can be nested inside other columns.

CommonUsageStats

Common statistics on the entry’s usage.

Contacts

Contact people for the entry.

CreateEntryGroupRequest

Request message for [CreateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.CreateEntryGroup].

CreateEntryRequest

Request message for [CreateEntry][google.cloud.datacatalog.v1.DataCatalog.CreateEntry].

CreatePolicyTagRequest

Request message for [CreatePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.CreatePolicyTag].

CreateTagRequest

Request message for [CreateTag][google.cloud.datacatalog.v1.DataCatalog.CreateTag].

CreateTagTemplateFieldRequest

Request message for [CreateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplateField].

CreateTagTemplateRequest

Request message for [CreateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplate].

CreateTaxonomyRequest

Request message for [CreateTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.CreateTaxonomy].

CrossRegionalSource

Cross-regional source used to import an existing taxonomy into a different region.

DataSource

Physical location of an entry.

DataSourceConnectionSpec

Specification that applies to a data source connection. Valid only for entries with the DATA_SOURCE_CONNECTION type. Only one of internal specs can be set at the time, and cannot be changed later.

DatabaseTableSpec

Specification that applies to a table resource. Valid only for entries with the TABLE type.

DataplexExternalTable

External table registered by Dataplex. Dataplex publishes data discovered from an asset into multiple other systems (BigQuery, DPMS) in form of tables. We call them “external tables”. External tables are also synced into the Data Catalog. This message contains pointers to those external tables (fully qualified name, resource name et cetera) within the Data Catalog.

DataplexFilesetSpec

Entry specyfication for a Dataplex fileset.

DataplexSpec

Common Dataplex fields.

DataplexTableSpec

Entry specification for a Dataplex table.

DatasetSpec

Specification that applies to a dataset. Valid only for entries with the DATASET type.

DeleteEntryGroupRequest

Request message for [DeleteEntryGroup][google.cloud.datacatalog.v1.DataCatalog.DeleteEntryGroup].

DeleteEntryRequest

Request message for [DeleteEntry][google.cloud.datacatalog.v1.DataCatalog.DeleteEntry].

DeletePolicyTagRequest

Request message for [DeletePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.DeletePolicyTag].

DeleteTagRequest

Request message for [DeleteTag][google.cloud.datacatalog.v1.DataCatalog.DeleteTag].

DeleteTagTemplateFieldRequest

Request message for [DeleteTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplateField].

DeleteTagTemplateRequest

Request message for [DeleteTagTemplate][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplate].

DeleteTaxonomyRequest

Request message for [DeleteTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.DeleteTaxonomy].

DumpItem

Wrapper for any item that can be contained in the dump.

Entry

Entry metadata. A Data Catalog entry represents another resource in Google Cloud Platform (such as a BigQuery dataset or a Pub/Sub topic) or outside of it. You can use the linked_resource field in the entry resource to refer to the original resource ID of the source system.

EntryGroup

Entry group metadata.

EntryOverview

Entry overview fields for rich text descriptions of entries.

ExportTaxonomiesRequest

Request message for [ExportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ExportTaxonomies].

ExportTaxonomiesResponse

Response message for [ExportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ExportTaxonomies].

FieldType

FilesetSpec

Specification that applies to a fileset. Valid only for entries with the ‘FILESET’ type.

GcsFileSpec

Specification of a single file in Cloud Storage.

GcsFilesetSpec

Describes a Cloud Storage fileset entry.

GetEntryGroupRequest

Request message for [GetEntryGroup][google.cloud.datacatalog.v1.DataCatalog.GetEntryGroup].

GetEntryRequest

Request message for [GetEntry][google.cloud.datacatalog.v1.DataCatalog.GetEntry].

GetPolicyTagRequest

Request message for [GetPolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.GetPolicyTag].

GetTagTemplateRequest

Request message for [GetTagTemplate][google.cloud.datacatalog.v1.DataCatalog.GetTagTemplate].

GetTaxonomyRequest

Request message for [GetTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.GetTaxonomy].

ImportEntriesMetadata

Metadata message for [long-running operation][google.longrunning.Operation] returned by the [ImportEntries][google.cloud.datacatalog.v1.DataCatalog.ImportEntries].

ImportEntriesRequest

Request message for [ImportEntries][google.cloud.datacatalog.v1.DataCatalog.ImportEntries] method.

ImportEntriesResponse

Response message for [long-running operation][google.longrunning.Operation] returned by the [ImportEntries][google.cloud.datacatalog.v1.DataCatalog.ImportEntries].

ImportTaxonomiesRequest

Request message for [ImportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ImportTaxonomies].

ImportTaxonomiesResponse

Response message for [ImportTaxonomies][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ImportTaxonomies].

InlineSource

Inline source containing taxonomies to import.

ListEntriesRequest

Request message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntriesResponse

Response message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntryGroupsRequest

Request message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListEntryGroupsResponse

Response message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListPolicyTagsRequest

Request message for [ListPolicyTags][google.cloud.datacatalog.v1.PolicyTagManager.ListPolicyTags].

ListPolicyTagsResponse

Response message for [ListPolicyTags][google.cloud.datacatalog.v1.PolicyTagManager.ListPolicyTags].

ListTagsRequest

Request message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

ListTagsResponse

Response message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

ListTaxonomiesRequest

Request message for [ListTaxonomies][google.cloud.datacatalog.v1.PolicyTagManager.ListTaxonomies].

ListTaxonomiesResponse

Response message for [ListTaxonomies][google.cloud.datacatalog.v1.PolicyTagManager.ListTaxonomies].

LookerSystemSpec

Specification that applies to entries that are part LOOKER system (user_specified_type)

LookupEntryRequest

Request message for [LookupEntry][google.cloud.datacatalog.v1.DataCatalog.LookupEntry].

ModelSpec

Specification that applies to a model. Valid only for entries with the MODEL type.

ModifyEntryContactsRequest

Request message for [ModifyEntryContacts][google.cloud.datacatalog.v1.DataCatalog.ModifyEntryContacts].

ModifyEntryOverviewRequest

Request message for [ModifyEntryOverview][google.cloud.datacatalog.v1.DataCatalog.ModifyEntryOverview].

PersonalDetails

Entry metadata relevant only to the user and private to them.

PhysicalSchema

Native schema used by a resource represented as an entry. Used by query engines for deserializing and parsing source data.

PolicyTag

Denotes one policy tag in a taxonomy, for example, SSN.

ReconcileTagsMetadata

[Long-running operation][google.longrunning.Operation] metadata message returned by the [ReconcileTags][google.cloud.datacatalog.v1.DataCatalog.ReconcileTags].

ReconcileTagsRequest

Request message for [ReconcileTags][google.cloud.datacatalog.v1.DataCatalog.ReconcileTags].

ReconcileTagsResponse

[Long-running operation][google.longrunning.Operation] response message returned by [ReconcileTags][google.cloud.datacatalog.v1.DataCatalog.ReconcileTags].

RenameTagTemplateFieldEnumValueRequest

Request message for [RenameTagTemplateFieldEnumValue][google.cloud.datacatalog.v1.DataCatalog.RenameTagTemplateFieldEnumValue].

RenameTagTemplateFieldRequest

Request message for [RenameTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.RenameTagTemplateField].

ReplaceTaxonomyRequest

Request message for [ReplaceTaxonomy][google.cloud.datacatalog.v1.PolicyTagManagerSerialization.ReplaceTaxonomy].

RoutineSpec

Specification that applies to a routine. Valid only for entries with the ROUTINE type.

Schema

Represents a schema, for example, a BigQuery, GoogleSQL, or Avro schema.

SearchCatalogRequest

Request message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResponse

Response message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResult

Result in the response to a search request.

SerializedPolicyTag

A nested protocol buffer that represents a policy tag and all its descendants.

SerializedTaxonomy

A nested protocol buffer that represents a taxonomy and the hierarchy of its policy tags. Used for taxonomy replacement, import, and export.

ServiceSpec

Specification that applies to a Service resource. Valid only for entries with the SERVICE type.

SqlDatabaseSystemSpec

Specification that applies to entries that are part SQL_DATABASE system (user_specified_type)

StarEntryRequest

Request message for [StarEntry][google.cloud.datacatalog.v1.DataCatalog.StarEntry].

StarEntryResponse

Response message for [StarEntry][google.cloud.datacatalog.v1.DataCatalog.StarEntry]. Empty for now

StorageProperties

Details the properties of the underlying storage.

SystemTimestamps

Timestamps associated with this resource in a particular system.

TableSpec

Normal BigQuery table specification.

Tag

Tags contain custom metadata and are attached to Data Catalog resources. Tags conform with the specification of their tag template.

TagField

Contains the value and additional information on a field within a [Tag][google.cloud.datacatalog.v1.Tag].

TagTemplate

A tag template defines a tag that can have one or more typed fields.

TagTemplateField

The template for an individual field within a tag template.

TaggedEntry

Wrapper containing Entry and information about Tags that should and should not be attached to it.

Taxonomy

A taxonomy is a collection of hierarchical policy tags that classify data along a common axis.

UnstarEntryRequest

Request message for [UnstarEntry][google.cloud.datacatalog.v1.DataCatalog.UnstarEntry].

UnstarEntryResponse

Response message for [UnstarEntry][google.cloud.datacatalog.v1.DataCatalog.UnstarEntry]. Empty for now

UpdateEntryGroupRequest

Request message for [UpdateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.UpdateEntryGroup].

UpdateEntryRequest

Request message for [UpdateEntry][google.cloud.datacatalog.v1.DataCatalog.UpdateEntry].

UpdatePolicyTagRequest

Request message for [UpdatePolicyTag][google.cloud.datacatalog.v1.PolicyTagManager.UpdatePolicyTag].

UpdateTagRequest

Request message for [UpdateTag][google.cloud.datacatalog.v1.DataCatalog.UpdateTag].

UpdateTagTemplateFieldRequest

Request message for [UpdateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplateField].

UpdateTagTemplateRequest

Request message for [UpdateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplate].

UpdateTaxonomyRequest

Request message for [UpdateTaxonomy][google.cloud.datacatalog.v1.PolicyTagManager.UpdateTaxonomy].

UsageSignal

The set of all usage signals that Data Catalog stores.

UsageStats

Detailed statistics on the entry’s usage.

VertexDatasetSpec

Specification for vertex dataset resources.

VertexModelSourceInfo

Detail description of the source information of a Vertex model.

VertexModelSpec

Specification for vertex model resources.

ViewSpec

Table view specification.

Enums

EntryType

Metadata automatically ingested from Google Cloud resources like BigQuery tables or Pub/Sub topics always uses enum values from EntryType as the type of entry.

IntegratedSystem

This enum lists all the systems that Data Catalog integrates with.

ManagingSystem

This enum describes all the systems that manage Taxonomy and PolicyTag resources in DataCatalog.

SearchResultType

The resource types that can be returned in search results.

TableSourceType

Table source type.