Struct google_api_proto::google::cloud::datacatalog::v1::data_catalog_client::DataCatalogClient

pub struct DataCatalogClient<T> { /* private fields */ }

Data Catalog API service allows you to discover, understand, and manage your data.

Implementations

impl<T> DataCatalogClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> DataCatalogClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn search_catalog( &mut self, request: impl IntoRequest<SearchCatalogRequest>, ) -> Result<Response<SearchCatalogResponse>, Status>

Searches Data Catalog for multiple resources like entries and tags that match a query.

This is a [Custom Method] (https://cloud.google.com/apis/design/custom_methods) that doesn’t return all information on a resource, only its ID and high level fields. To get more information, you can subsequently call specific get methods.

Note: Data Catalog search queries don’t guarantee full recall. Results that match your query might not be returned, even in subsequent result pages. Additionally, returned (and not returned) results can vary if you repeat search queries.

For more information, see [Data Catalog search syntax] (https://cloud.google.com/data-catalog/docs/how-to/search-reference).

pub async fn create_entry_group( &mut self, request: impl IntoRequest<CreateEntryGroupRequest>, ) -> Result<Response<EntryGroup>, Status>

Creates an entry group.

An entry group contains logically related entries together with Cloud Identity and Access Management policies. These policies specify users who can create, edit, and view entries within entry groups.

Data Catalog automatically creates entry groups with names that start with the @ symbol for the following resources:

BigQuery entries (@bigquery)

Pub/Sub topics (@pubsub)

Dataproc Metastore services (@dataproc_metastore_{SERVICE_NAME_HASH})

You can create your own entry groups for Cloud Storage fileset entries and custom entries together with the corresponding IAM policies. User-created entry groups can’t contain the @ symbol, it is reserved for automatically created groups.

Entry groups, like entries, can be searched.

A maximum of 10,000 entry groups may be created per organization across all locations.

You must enable the Data Catalog API in the project identified by the parent parameter. For more information, see Data Catalog resource project.

pub async fn get_entry_group( &mut self, request: impl IntoRequest<GetEntryGroupRequest>, ) -> Result<Response<EntryGroup>, Status>

Gets an entry group.

pub async fn update_entry_group( &mut self, request: impl IntoRequest<UpdateEntryGroupRequest>, ) -> Result<Response<EntryGroup>, Status>

Updates an entry group.

You must enable the Data Catalog API in the project identified by the entry_group.name parameter. For more information, see Data Catalog resource project.

pub async fn delete_entry_group( &mut self, request: impl IntoRequest<DeleteEntryGroupRequest>, ) -> Result<Response<()>, Status>

Deletes an entry group.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see Data Catalog resource project.

pub async fn list_entry_groups( &mut self, request: impl IntoRequest<ListEntryGroupsRequest>, ) -> Result<Response<ListEntryGroupsResponse>, Status>

Lists entry groups.

pub async fn create_entry( &mut self, request: impl IntoRequest<CreateEntryRequest>, ) -> Result<Response<Entry>, Status>

Creates an entry.

You can create entries only with ‘FILESET’, ‘CLUSTER’, ‘DATA_STREAM’, or custom types. Data Catalog automatically creates entries with other types during metadata ingestion from integrated systems.

You must enable the Data Catalog API in the project identified by the parent parameter. For more information, see Data Catalog resource project.

An entry group can have a maximum of 100,000 entries.

pub async fn update_entry( &mut self, request: impl IntoRequest<UpdateEntryRequest>, ) -> Result<Response<Entry>, Status>

Updates an existing entry.

You must enable the Data Catalog API in the project identified by the entry.name parameter. For more information, see Data Catalog resource project.

pub async fn delete_entry( &mut self, request: impl IntoRequest<DeleteEntryRequest>, ) -> Result<Response<()>, Status>

Deletes an existing entry.

You can delete only the entries created by the [CreateEntry][google.cloud.datacatalog.v1.DataCatalog.CreateEntry] method.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see Data Catalog resource project.

pub async fn get_entry( &mut self, request: impl IntoRequest<GetEntryRequest>, ) -> Result<Response<Entry>, Status>

Gets an entry.

pub async fn lookup_entry( &mut self, request: impl IntoRequest<LookupEntryRequest>, ) -> Result<Response<Entry>, Status>

Gets an entry by its target resource name.

The resource name comes from the source Google Cloud Platform service.

pub async fn list_entries( &mut self, request: impl IntoRequest<ListEntriesRequest>, ) -> Result<Response<ListEntriesResponse>, Status>

Lists entries.

Note: Currently, this method can list only custom entries. To get a list of both custom and automatically created entries, use [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

pub async fn modify_entry_overview( &mut self, request: impl IntoRequest<ModifyEntryOverviewRequest>, ) -> Result<Response<EntryOverview>, Status>

Modifies entry overview, part of the business context of an [Entry][google.cloud.datacatalog.v1.Entry].

To call this method, you must have the datacatalog.entries.updateOverview IAM permission on the corresponding project.

pub async fn modify_entry_contacts( &mut self, request: impl IntoRequest<ModifyEntryContactsRequest>, ) -> Result<Response<Contacts>, Status>

Modifies contacts, part of the business context of an [Entry][google.cloud.datacatalog.v1.Entry].

To call this method, you must have the datacatalog.entries.updateContacts IAM permission on the corresponding project.

pub async fn create_tag_template( &mut self, request: impl IntoRequest<CreateTagTemplateRequest>, ) -> Result<Response<TagTemplate>, Status>

Creates a tag template.

You must enable the Data Catalog API in the project identified by the parent parameter. For more information, see [Data Catalog resource project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project).

pub async fn get_tag_template( &mut self, request: impl IntoRequest<GetTagTemplateRequest>, ) -> Result<Response<TagTemplate>, Status>

Gets a tag template.

pub async fn update_tag_template( &mut self, request: impl IntoRequest<UpdateTagTemplateRequest>, ) -> Result<Response<TagTemplate>, Status>

Updates a tag template.

You can’t update template fields with this method. These fields are separate resources with their own create, update, and delete methods.

You must enable the Data Catalog API in the project identified by the tag_template.name parameter. For more information, see Data Catalog resource project.

pub async fn delete_tag_template( &mut self, request: impl IntoRequest<DeleteTagTemplateRequest>, ) -> Result<Response<()>, Status>

Deletes a tag template and all tags that use it.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see Data Catalog resource project.

pub async fn create_tag_template_field( &mut self, request: impl IntoRequest<CreateTagTemplateFieldRequest>, ) -> Result<Response<TagTemplateField>, Status>

Creates a field in a tag template.

You must enable the Data Catalog API in the project identified by the parent parameter. For more information, see Data Catalog resource project.

pub async fn update_tag_template_field( &mut self, request: impl IntoRequest<UpdateTagTemplateFieldRequest>, ) -> Result<Response<TagTemplateField>, Status>

Updates a field in a tag template.

You can’t update the field type with this method.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see Data Catalog resource project.

pub async fn rename_tag_template_field( &mut self, request: impl IntoRequest<RenameTagTemplateFieldRequest>, ) -> Result<Response<TagTemplateField>, Status>

Renames a field in a tag template.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see [Data Catalog resource project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project).

pub async fn rename_tag_template_field_enum_value( &mut self, request: impl IntoRequest<RenameTagTemplateFieldEnumValueRequest>, ) -> Result<Response<TagTemplateField>, Status>

Renames an enum value in a tag template.

Within a single enum field, enum values must be unique.

pub async fn delete_tag_template_field( &mut self, request: impl IntoRequest<DeleteTagTemplateFieldRequest>, ) -> Result<Response<()>, Status>

Deletes a field in a tag template and all uses of this field from the tags based on this template.

You must enable the Data Catalog API in the project identified by the name parameter. For more information, see Data Catalog resource project.

pub async fn create_tag( &mut self, request: impl IntoRequest<CreateTagRequest>, ) -> Result<Response<Tag>, Status>

Creates a tag and assigns it to:

An [Entry][google.cloud.datacatalog.v1.Entry] if the method name is projects.locations.entryGroups.entries.tags.create.

Or [EntryGroup][google.cloud.datacatalog.v1.EntryGroup]if the method name is projects.locations.entryGroups.tags.create.

Note: The project identified by the parent parameter for the [tag] (https://cloud.google.com/data-catalog/docs/reference/rest/v1/projects.locations.entryGroups.entries.tags/create#path-parameters) and the [tag template] (https://cloud.google.com/data-catalog/docs/reference/rest/v1/projects.locations.tagTemplates/create#path-parameters) used to create the tag must be in the same organization.

pub async fn update_tag( &mut self, request: impl IntoRequest<UpdateTagRequest>, ) -> Result<Response<Tag>, Status>

Updates an existing tag.

pub async fn delete_tag( &mut self, request: impl IntoRequest<DeleteTagRequest>, ) -> Result<Response<()>, Status>

Deletes a tag.

pub async fn list_tags( &mut self, request: impl IntoRequest<ListTagsRequest>, ) -> Result<Response<ListTagsResponse>, Status>

Lists tags assigned to an [Entry][google.cloud.datacatalog.v1.Entry]. The [columns][google.cloud.datacatalog.v1.Tag.column] in the response are lowercased.

pub async fn reconcile_tags( &mut self, request: impl IntoRequest<ReconcileTagsRequest>, ) -> Result<Response<Operation>, Status>

ReconcileTags creates or updates a list of tags on the entry. If the [ReconcileTagsRequest.force_delete_missing][google.cloud.datacatalog.v1.ReconcileTagsRequest.force_delete_missing] parameter is set, the operation deletes tags not included in the input tag list.

ReconcileTags returns a [long-running operation] [google.longrunning.Operation] resource that can be queried with [Operations.GetOperation][google.longrunning.Operations.GetOperation] to return [ReconcileTagsMetadata] [google.cloud.datacatalog.v1.ReconcileTagsMetadata] and a [ReconcileTagsResponse] [google.cloud.datacatalog.v1.ReconcileTagsResponse] message.

pub async fn star_entry( &mut self, request: impl IntoRequest<StarEntryRequest>, ) -> Result<Response<StarEntryResponse>, Status>

Marks an [Entry][google.cloud.datacatalog.v1.Entry] as starred by the current user. Starring information is private to each user.

pub async fn unstar_entry( &mut self, request: impl IntoRequest<UnstarEntryRequest>, ) -> Result<Response<UnstarEntryResponse>, Status>

Marks an [Entry][google.cloud.datacatalog.v1.Entry] as NOT starred by the current user. Starring information is private to each user.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Sets an access control policy for a resource. Replaces any existing policy.

Supported resources are:

Tag templates

Entry groups

Note: This method sets policies only within Data Catalog and can’t be used to manage policies in BigQuery, Pub/Sub, Dataproc Metastore, and any external Google Cloud Platform resources synced with the Data Catalog.

To call this method, you must have the following Google IAM permissions:

datacatalog.tagTemplates.setIamPolicy to set policies on tag templates.

datacatalog.entryGroups.setIamPolicy to set policies on entry groups.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets the access control policy for a resource.

May return:

ANOT_FOUND error if the resource doesn’t exist or you don’t have the permission to view it.

An empty policy if the resource exists but doesn’t have a set policy.

Supported resources are:

Tag templates

Entry groups

Note: This method doesn’t get policies from Google Cloud Platform resources ingested into Data Catalog.

To call this method, you must have the following Google IAM permissions:

datacatalog.tagTemplates.getIamPolicy to get policies on tag templates.

datacatalog.entryGroups.getIamPolicy to get policies on entry groups.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest>, ) -> Result<Response<TestIamPermissionsResponse>, Status>

Gets your permissions on a resource.

Returns an empty set of permissions if the resource doesn’t exist.

Supported resources are:

Tag templates

Entry groups

Note: This method gets policies only within Data Catalog and can’t be used to get policies from BigQuery, Pub/Sub, Dataproc Metastore, and any external Google Cloud Platform resources ingested into Data Catalog.

No Google IAM permissions are required to call this method.

pub async fn import_entries( &mut self, request: impl IntoRequest<ImportEntriesRequest>, ) -> Result<Response<Operation>, Status>

Imports entries from a source, such as data previously dumped into a Cloud Storage bucket, into Data Catalog. Import of entries is a sync operation that reconciles the state of the third-party system with the Data Catalog.

ImportEntries accepts source data snapshots of a third-party system. Snapshot should be delivered as a .wire or base65-encoded .txt file containing a sequence of Protocol Buffer messages of [DumpItem][google.cloud.datacatalog.v1.DumpItem] type.

ImportEntries returns a [long-running operation] [google.longrunning.Operation] resource that can be queried with [Operations.GetOperation][google.longrunning.Operations.GetOperation] to return [ImportEntriesMetadata][google.cloud.datacatalog.v1.ImportEntriesMetadata] and an [ImportEntriesResponse][google.cloud.datacatalog.v1.ImportEntriesResponse] message.