Struct google_api_proto::google::cloud::datacatalog::v1::policy_tag_manager_client::PolicyTagManagerClient

pub struct PolicyTagManagerClient<T> { /* private fields */ }

Policy Tag Manager API service allows you to manage your policy tags and taxonomies.

Policy tags are used to tag BigQuery columns and apply additional access control policies. A taxonomy is a hierarchical grouping of policy tags that classify data along a common axis.

Implementations

impl<T> PolicyTagManagerClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> PolicyTagManagerClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_taxonomy( &mut self, request: impl IntoRequest<CreateTaxonomyRequest> ) -> Result<Response<Taxonomy>, Status>

Creates a taxonomy in a specified project.

The taxonomy is initially empty, that is, it doesn’t contain policy tags.

pub async fn delete_taxonomy( &mut self, request: impl IntoRequest<DeleteTaxonomyRequest> ) -> Result<Response<()>, Status>

Deletes a taxonomy, including all policy tags in this taxonomy, their associated policies, and the policy tags references from BigQuery columns.

pub async fn update_taxonomy( &mut self, request: impl IntoRequest<UpdateTaxonomyRequest> ) -> Result<Response<Taxonomy>, Status>

Updates a taxonomy, including its display name, description, and activated policy types.

pub async fn list_taxonomies( &mut self, request: impl IntoRequest<ListTaxonomiesRequest> ) -> Result<Response<ListTaxonomiesResponse>, Status>

Lists all taxonomies in a project in a particular location that you have a permission to view.

pub async fn get_taxonomy( &mut self, request: impl IntoRequest<GetTaxonomyRequest> ) -> Result<Response<Taxonomy>, Status>

Gets a taxonomy.

pub async fn create_policy_tag( &mut self, request: impl IntoRequest<CreatePolicyTagRequest> ) -> Result<Response<PolicyTag>, Status>

Creates a policy tag in a taxonomy.

pub async fn delete_policy_tag( &mut self, request: impl IntoRequest<DeletePolicyTagRequest> ) -> Result<Response<()>, Status>

Deletes a policy tag together with the following:

All of its descendant policy tags, if any

Policies associated with the policy tag and its descendants

References from BigQuery table schema of the policy tag and its descendants

pub async fn update_policy_tag( &mut self, request: impl IntoRequest<UpdatePolicyTagRequest> ) -> Result<Response<PolicyTag>, Status>

Updates a policy tag, including its display name, description, and parent policy tag.

pub async fn list_policy_tags( &mut self, request: impl IntoRequest<ListPolicyTagsRequest> ) -> Result<Response<ListPolicyTagsResponse>, Status>

Lists all policy tags in a taxonomy.

pub async fn get_policy_tag( &mut self, request: impl IntoRequest<GetPolicyTagRequest> ) -> Result<Response<PolicyTag>, Status>

Gets a policy tag.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Gets the IAM policy for a policy tag or a taxonomy.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Sets the IAM policy for a policy tag or a taxonomy.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest> ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns your permissions on a specified policy tag or taxonomy.