Struct google_api_proto::google::cloud::dataform::v1beta1::dataform_client::DataformClient

pub struct DataformClient<T> { /* private fields */ }

Dataform is a service to develop, create, document, test, and update curated tables in BigQuery.

Implementations

impl<T> DataformClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> DataformClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_repositories( &mut self, request: impl IntoRequest<ListRepositoriesRequest>, ) -> Result<Response<ListRepositoriesResponse>, Status>

Lists Repositories in a given project and location.

pub async fn get_repository( &mut self, request: impl IntoRequest<GetRepositoryRequest>, ) -> Result<Response<Repository>, Status>

Fetches a single Repository.

pub async fn create_repository( &mut self, request: impl IntoRequest<CreateRepositoryRequest>, ) -> Result<Response<Repository>, Status>

Creates a new Repository in a given project and location.

pub async fn update_repository( &mut self, request: impl IntoRequest<UpdateRepositoryRequest>, ) -> Result<Response<Repository>, Status>

Updates a single Repository.

pub async fn delete_repository( &mut self, request: impl IntoRequest<DeleteRepositoryRequest>, ) -> Result<Response<()>, Status>

Deletes a single Repository.

pub async fn commit_repository_changes( &mut self, request: impl IntoRequest<CommitRepositoryChangesRequest>, ) -> Result<Response<()>, Status>

Applies a Git commit to a Repository. The Repository must not have a value for git_remote_settings.url.

pub async fn read_repository_file( &mut self, request: impl IntoRequest<ReadRepositoryFileRequest>, ) -> Result<Response<ReadRepositoryFileResponse>, Status>

Returns the contents of a file (inside a Repository). The Repository must not have a value for git_remote_settings.url.

pub async fn query_repository_directory_contents( &mut self, request: impl IntoRequest<QueryRepositoryDirectoryContentsRequest>, ) -> Result<Response<QueryRepositoryDirectoryContentsResponse>, Status>

Returns the contents of a given Repository directory. The Repository must not have a value for git_remote_settings.url.

pub async fn fetch_repository_history( &mut self, request: impl IntoRequest<FetchRepositoryHistoryRequest>, ) -> Result<Response<FetchRepositoryHistoryResponse>, Status>

Fetches a Repository’s history of commits. The Repository must not have a value for git_remote_settings.url.

pub async fn compute_repository_access_token_status( &mut self, request: impl IntoRequest<ComputeRepositoryAccessTokenStatusRequest>, ) -> Result<Response<ComputeRepositoryAccessTokenStatusResponse>, Status>

Computes a Repository’s Git access token status.

pub async fn fetch_remote_branches( &mut self, request: impl IntoRequest<FetchRemoteBranchesRequest>, ) -> Result<Response<FetchRemoteBranchesResponse>, Status>

Fetches a Repository’s remote branches.

pub async fn list_workspaces( &mut self, request: impl IntoRequest<ListWorkspacesRequest>, ) -> Result<Response<ListWorkspacesResponse>, Status>

Lists Workspaces in a given Repository.

pub async fn get_workspace( &mut self, request: impl IntoRequest<GetWorkspaceRequest>, ) -> Result<Response<Workspace>, Status>

Fetches a single Workspace.

pub async fn create_workspace( &mut self, request: impl IntoRequest<CreateWorkspaceRequest>, ) -> Result<Response<Workspace>, Status>

Creates a new Workspace in a given Repository.

pub async fn delete_workspace( &mut self, request: impl IntoRequest<DeleteWorkspaceRequest>, ) -> Result<Response<()>, Status>

Deletes a single Workspace.

pub async fn install_npm_packages( &mut self, request: impl IntoRequest<InstallNpmPackagesRequest>, ) -> Result<Response<InstallNpmPackagesResponse>, Status>

Installs dependency NPM packages (inside a Workspace).

pub async fn pull_git_commits( &mut self, request: impl IntoRequest<PullGitCommitsRequest>, ) -> Result<Response<()>, Status>

Pulls Git commits from the Repository’s remote into a Workspace.

pub async fn push_git_commits( &mut self, request: impl IntoRequest<PushGitCommitsRequest>, ) -> Result<Response<()>, Status>

Pushes Git commits from a Workspace to the Repository’s remote.

pub async fn fetch_file_git_statuses( &mut self, request: impl IntoRequest<FetchFileGitStatusesRequest>, ) -> Result<Response<FetchFileGitStatusesResponse>, Status>

Fetches Git statuses for the files in a Workspace.

pub async fn fetch_git_ahead_behind( &mut self, request: impl IntoRequest<FetchGitAheadBehindRequest>, ) -> Result<Response<FetchGitAheadBehindResponse>, Status>

Fetches Git ahead/behind against a remote branch.

pub async fn commit_workspace_changes( &mut self, request: impl IntoRequest<CommitWorkspaceChangesRequest>, ) -> Result<Response<()>, Status>

Applies a Git commit for uncommitted files in a Workspace.

pub async fn reset_workspace_changes( &mut self, request: impl IntoRequest<ResetWorkspaceChangesRequest>, ) -> Result<Response<()>, Status>

Performs a Git reset for uncommitted files in a Workspace.

pub async fn fetch_file_diff( &mut self, request: impl IntoRequest<FetchFileDiffRequest>, ) -> Result<Response<FetchFileDiffResponse>, Status>

Fetches Git diff for an uncommitted file in a Workspace.

pub async fn query_directory_contents( &mut self, request: impl IntoRequest<QueryDirectoryContentsRequest>, ) -> Result<Response<QueryDirectoryContentsResponse>, Status>

Returns the contents of a given Workspace directory.

pub async fn make_directory( &mut self, request: impl IntoRequest<MakeDirectoryRequest>, ) -> Result<Response<MakeDirectoryResponse>, Status>

Creates a directory inside a Workspace.

pub async fn remove_directory( &mut self, request: impl IntoRequest<RemoveDirectoryRequest>, ) -> Result<Response<()>, Status>

Deletes a directory (inside a Workspace) and all of its contents.

pub async fn move_directory( &mut self, request: impl IntoRequest<MoveDirectoryRequest>, ) -> Result<Response<MoveDirectoryResponse>, Status>

Moves a directory (inside a Workspace), and all of its contents, to a new location.

pub async fn read_file( &mut self, request: impl IntoRequest<ReadFileRequest>, ) -> Result<Response<ReadFileResponse>, Status>

Returns the contents of a file (inside a Workspace).

pub async fn remove_file( &mut self, request: impl IntoRequest<RemoveFileRequest>, ) -> Result<Response<()>, Status>

Deletes a file (inside a Workspace).

pub async fn move_file( &mut self, request: impl IntoRequest<MoveFileRequest>, ) -> Result<Response<MoveFileResponse>, Status>

Moves a file (inside a Workspace) to a new location.

pub async fn write_file( &mut self, request: impl IntoRequest<WriteFileRequest>, ) -> Result<Response<WriteFileResponse>, Status>

Writes to a file (inside a Workspace).

pub async fn list_release_configs( &mut self, request: impl IntoRequest<ListReleaseConfigsRequest>, ) -> Result<Response<ListReleaseConfigsResponse>, Status>

Lists ReleaseConfigs in a given Repository.

pub async fn get_release_config( &mut self, request: impl IntoRequest<GetReleaseConfigRequest>, ) -> Result<Response<ReleaseConfig>, Status>

Fetches a single ReleaseConfig.

pub async fn create_release_config( &mut self, request: impl IntoRequest<CreateReleaseConfigRequest>, ) -> Result<Response<ReleaseConfig>, Status>

Creates a new ReleaseConfig in a given Repository.

pub async fn update_release_config( &mut self, request: impl IntoRequest<UpdateReleaseConfigRequest>, ) -> Result<Response<ReleaseConfig>, Status>

Updates a single ReleaseConfig.

pub async fn delete_release_config( &mut self, request: impl IntoRequest<DeleteReleaseConfigRequest>, ) -> Result<Response<()>, Status>

Deletes a single ReleaseConfig.

pub async fn list_compilation_results( &mut self, request: impl IntoRequest<ListCompilationResultsRequest>, ) -> Result<Response<ListCompilationResultsResponse>, Status>

Lists CompilationResults in a given Repository.

pub async fn get_compilation_result( &mut self, request: impl IntoRequest<GetCompilationResultRequest>, ) -> Result<Response<CompilationResult>, Status>

Fetches a single CompilationResult.

pub async fn create_compilation_result( &mut self, request: impl IntoRequest<CreateCompilationResultRequest>, ) -> Result<Response<CompilationResult>, Status>

Creates a new CompilationResult in a given project and location.

pub async fn query_compilation_result_actions( &mut self, request: impl IntoRequest<QueryCompilationResultActionsRequest>, ) -> Result<Response<QueryCompilationResultActionsResponse>, Status>

Returns CompilationResultActions in a given CompilationResult.

pub async fn list_workflow_configs( &mut self, request: impl IntoRequest<ListWorkflowConfigsRequest>, ) -> Result<Response<ListWorkflowConfigsResponse>, Status>

Lists WorkflowConfigs in a given Repository.

pub async fn get_workflow_config( &mut self, request: impl IntoRequest<GetWorkflowConfigRequest>, ) -> Result<Response<WorkflowConfig>, Status>

Fetches a single WorkflowConfig.

pub async fn create_workflow_config( &mut self, request: impl IntoRequest<CreateWorkflowConfigRequest>, ) -> Result<Response<WorkflowConfig>, Status>

Creates a new WorkflowConfig in a given Repository.

pub async fn update_workflow_config( &mut self, request: impl IntoRequest<UpdateWorkflowConfigRequest>, ) -> Result<Response<WorkflowConfig>, Status>

Updates a single WorkflowConfig.

pub async fn delete_workflow_config( &mut self, request: impl IntoRequest<DeleteWorkflowConfigRequest>, ) -> Result<Response<()>, Status>

Deletes a single WorkflowConfig.

pub async fn list_workflow_invocations( &mut self, request: impl IntoRequest<ListWorkflowInvocationsRequest>, ) -> Result<Response<ListWorkflowInvocationsResponse>, Status>

Lists WorkflowInvocations in a given Repository.

pub async fn get_workflow_invocation( &mut self, request: impl IntoRequest<GetWorkflowInvocationRequest>, ) -> Result<Response<WorkflowInvocation>, Status>

Fetches a single WorkflowInvocation.

pub async fn create_workflow_invocation( &mut self, request: impl IntoRequest<CreateWorkflowInvocationRequest>, ) -> Result<Response<WorkflowInvocation>, Status>

Creates a new WorkflowInvocation in a given Repository.

pub async fn delete_workflow_invocation( &mut self, request: impl IntoRequest<DeleteWorkflowInvocationRequest>, ) -> Result<Response<()>, Status>

Deletes a single WorkflowInvocation.

pub async fn cancel_workflow_invocation( &mut self, request: impl IntoRequest<CancelWorkflowInvocationRequest>, ) -> Result<Response<()>, Status>

Requests cancellation of a running WorkflowInvocation.

pub async fn query_workflow_invocation_actions( &mut self, request: impl IntoRequest<QueryWorkflowInvocationActionsRequest>, ) -> Result<Response<QueryWorkflowInvocationActionsResponse>, Status>

Returns WorkflowInvocationActions in a given WorkflowInvocation.