Module google_api_proto::google::cloud::datafusion::v1

Modules

accelerator

Nested message and enum types in Accelerator.

data_fusion_client

Generated client implementations.

instance

Nested message and enum types in Instance.

version

Nested message and enum types in Version.

Structs

Accelerator

Identifies Data Fusion accelerators for an instance.

CreateInstanceRequest

Request message for creating a Data Fusion instance.

CryptoKeyConfig

The crypto key configuration. This field is used by the Customer-managed encryption keys (CMEK) feature.

DeleteInstanceRequest

Request message for deleting a Data Fusion instance.

GetInstanceRequest

Request message for getting details about a Data Fusion instance.

Instance

Represents a Data Fusion instance.

ListAvailableVersionsRequest

Request message for the list available versions request.

ListAvailableVersionsResponse

Response message for the list available versions request.

ListInstancesRequest

Request message for listing Data Fusion instances.

ListInstancesResponse

Response message for the list instance request.

NetworkConfig

Network configuration for a Data Fusion instance. These configurations are used for peering with the customer network. Configurations are optional when a public Data Fusion instance is to be created. However, providing these configurations allows several benefits, such as reduced network latency while accessing the customer resources from managed Data Fusion instance nodes, as well as access to the customer on-prem resources.

OperationMetadata

Represents the metadata of a long-running operation.

RestartInstanceRequest

Request message for restarting a Data Fusion instance.

UpdateInstanceRequest

Request message for updating a Data Fusion instance. Data Fusion allows updating the labels, options, and stack driver settings. This is also used for CDF version upgrade.

Version

The Data Fusion version. This proto message stores information about certain Data Fusion version, which is used for Data Fusion version upgrade.