Module google_api_proto::google::cloud::dataplex::v1

Modules

action

Nested message and enum types in Action.

aspect_type

Nested message and enum types in AspectType.

asset

Nested message and enum types in Asset.

catalog_service_client

Generated client implementations.

content

Nested message and enum types in Content.

content_service_client

Generated client implementations.

data_attribute_binding

Nested message and enum types in DataAttributeBinding.

data_profile_result

Nested message and enum types in DataProfileResult.

data_profile_spec

Nested message and enum types in DataProfileSpec.

data_quality_result

Nested message and enum types in DataQualityResult.

data_quality_rule

Nested message and enum types in DataQualityRule.

data_quality_scan_rule_result

Nested message and enum types in DataQualityScanRuleResult.

data_quality_spec

Nested message and enum types in DataQualitySpec.

data_scan

Nested message and enum types in DataScan.

data_scan_event

Nested message and enum types in DataScanEvent.

data_scan_job

Nested message and enum types in DataScanJob.

data_scan_service_client

Generated client implementations.

data_source

Nested message and enum types in DataSource.

data_taxonomy_service_client

Generated client implementations.

dataplex_service_client

Generated client implementations.

discovery_event

Nested message and enum types in DiscoveryEvent.

entity

Nested message and enum types in Entity.

entry_source

Nested message and enum types in EntrySource.

entry_type

Nested message and enum types in EntryType.

environment

Nested message and enum types in Environment.

get_content_request

Nested message and enum types in GetContentRequest.

get_data_scan_job_request

Nested message and enum types in GetDataScanJobRequest.

get_data_scan_request

Nested message and enum types in GetDataScanRequest.

get_entity_request

Nested message and enum types in GetEntityRequest.

governance_event

Nested message and enum types in GovernanceEvent.

job

Nested message and enum types in Job.

job_event

Nested message and enum types in JobEvent.

lake

Nested message and enum types in Lake.

list_entities_request

Nested message and enum types in ListEntitiesRequest.

metadata_service_client

Generated client implementations.

scanned_data

Nested message and enum types in ScannedData.

schema

Nested message and enum types in Schema.

search_entries_result

Nested message and enum types in SearchEntriesResult.

session_event

Nested message and enum types in SessionEvent.

storage_access

Nested message and enum types in StorageAccess.

storage_format

Nested message and enum types in StorageFormat.

task

Nested message and enum types in Task.

trigger

Nested message and enum types in Trigger.

zone

Nested message and enum types in Zone.

Structs

Action

Action represents an issue requiring administrator action for resolution.

Aspect

An aspect is a single piece of metadata describing an entry.

AspectSource

AspectSource contains source system related information for the aspect.

AspectType

Aspect Type is a template for creating Aspects, and represents the JSON-schema for a given Entry, e.g., BigQuery Table Schema.

Asset

An asset represents a cloud resource that is being managed within a lake as a member of a zone.

AssetStatus

Aggregated status of the underlying assets of a lake or zone.

CancelJobRequest

Cancel task jobs.

Content

Content represents a user-visible notebook or a sql script

CreateAspectTypeRequest

Create AspectType Request

CreateAssetRequest

Create asset request.

CreateContentRequest

Create content request.

CreateDataAttributeBindingRequest

Create DataAttributeBinding request.

CreateDataAttributeRequest

Create DataAttribute request.

CreateDataScanRequest

Create dataScan request.

CreateDataTaxonomyRequest

Create DataTaxonomy request.

CreateEntityRequest

Create a metadata entity request.

CreateEntryGroupRequest

Create EntryGroup Request

CreateEntryRequest

CreateEntryTypeRequest

Create EntryType Request

CreateEnvironmentRequest

Create environment request.

CreateLakeRequest

Create lake request.

CreatePartitionRequest

Create metadata partition request.

CreateTaskRequest

Create task request.

CreateZoneRequest

Create zone request.

DataAccessSpec

DataAccessSpec holds the access control configuration to be enforced on data stored within resources (eg: rows, columns in BigQuery Tables). When associated with data, the data is only accessible to principals explicitly granted access through the DataAccessSpec. Principals with access to the containing resource are not implicitly granted access.

DataAttribute

Denotes one dataAttribute in a dataTaxonomy, for example, PII. DataAttribute resources can be defined in a hierarchy. A single dataAttribute resource can contain specs of multiple types

DataAttributeBinding

DataAttributeBinding represents binding of attributes to resources. Eg: Bind ‘CustomerInfo’ entity with ‘PII’ attribute.

DataProfileResult

DataProfileResult defines the output of DataProfileScan. Each field of the table will have field type specific profile result.

DataProfileSpec

DataProfileScan related setting.

DataQualityColumnResult

DataQualityColumnResult provides a more detailed, per-column view of the results.

DataQualityDimension

A dimension captures data quality intent about a defined subset of the rules specified.

DataQualityDimensionResult

DataQualityDimensionResult provides a more detailed, per-dimension view of the results.

DataQualityResult

The output of a DataQualityScan.

DataQualityRule

A rule captures data quality intent about a data source.

DataQualityRuleResult

DataQualityRuleResult provides a more detailed, per-rule view of the results.

DataQualityScanRuleResult

Information about the result of a data quality rule for data quality scan. The monitored resource is ‘DataScan’.

DataQualitySpec

DataQualityScan related setting.

DataScan

Represents a user-visible job which provides the insights for the related data source.

DataScanEvent

These messages contain information about the execution of a datascan. The monitored resource is ‘DataScan’ Next ID: 13

DataScanJob

A DataScanJob represents an instance of DataScan execution.

DataSource

The data source for DataScan.

DataTaxonomy

DataTaxonomy represents a set of hierarchical DataAttributes resources, grouped with a common theme Eg: ‘SensitiveDataTaxonomy’ can have attributes to manage PII data. It is defined at project level.

DeleteAspectTypeRequest

Delele AspectType Request

DeleteAssetRequest

Delete asset request.

DeleteContentRequest

Delete content request.

DeleteDataAttributeBindingRequest

Delete DataAttributeBinding request.

DeleteDataAttributeRequest

Delete DataAttribute request.

DeleteDataScanRequest

Delete dataScan request.

DeleteDataTaxonomyRequest

Delete DataTaxonomy request.

DeleteEntityRequest

Delete a metadata entity request.

DeleteEntryGroupRequest

Delele EntryGroup Request

DeleteEntryRequest

DeleteEntryTypeRequest

Delele EntryType Request

DeleteEnvironmentRequest

Delete environment request.

DeleteLakeRequest

Delete lake request.

DeletePartitionRequest

Delete metadata partition request.

DeleteTaskRequest

Delete task request.

DeleteZoneRequest

Delete zone request.

DiscoveryEvent

The payload associated with Discovery data processing.

Entity

Represents tables and fileset metadata contained within a zone.

Entry

An entry is a representation of a data asset which can be described by various metadata.

EntryGroup

An Entry Group represents a logical grouping of one or more Entries.

EntrySource

EntrySource contains source system related information for the entry.

EntryType

Entry Type is a template for creating Entries.

Environment

Environment represents a user-visible compute infrastructure for analytics within a lake.

GenerateDataQualityRulesRequest

Request details for generating data quality rule recommendations.

GenerateDataQualityRulesResponse

Response details for data quality rule recommendations.

GetAspectTypeRequest

Get AspectType request

GetAssetRequest

Get asset request.

GetContentRequest

Get content request.

GetDataAttributeBindingRequest

Get DataAttributeBinding request.

GetDataAttributeRequest

Get DataAttribute request.

GetDataScanJobRequest

Get DataScanJob request.

GetDataScanRequest

Get dataScan request.

GetDataTaxonomyRequest

Get DataTaxonomy request.

GetEntityRequest

Get metadata entity request.

GetEntryGroupRequest

Get EntryGroup request.

GetEntryRequest

GetEntryTypeRequest

Get EntryType request

GetEnvironmentRequest

Get environment request.

GetJobRequest

Get job request.

GetLakeRequest

Get lake request.

GetPartitionRequest

Get metadata partition request.

GetTaskRequest

Get task request.

GetZoneRequest

Get zone request.

GovernanceEvent

Payload associated with Governance related log events.

Job

A job represents an instance of a task.

JobEvent

The payload associated with Job logs that contains events describing jobs that have run within a Lake.

Lake

A lake is a centralized repository for managing enterprise data across the organization distributed across many cloud projects, and stored in a variety of storage services such as Google Cloud Storage and BigQuery. The resources attached to a lake are referred to as managed resources. Data within these managed resources can be structured or unstructured. A lake provides data admins with tools to organize, secure and manage their data at scale, and provides data scientists and data engineers an integrated experience to easily search, discover, analyze and transform data and associated metadata.

ListActionsResponse

List actions response.

ListAspectTypesRequest

List AspectTypes request

ListAspectTypesResponse

List AspectTypes response

ListAssetActionsRequest

List asset actions request.

ListAssetsRequest

List assets request.

ListAssetsResponse

List assets response.

ListContentRequest

List content request. Returns the BASIC Content view.

ListContentResponse

List content response.

ListDataAttributeBindingsRequest

List DataAttributeBindings request.

ListDataAttributeBindingsResponse

List DataAttributeBindings response.

ListDataAttributesRequest

List DataAttributes request.

ListDataAttributesResponse

List DataAttributes response.

ListDataScanJobsRequest

List DataScanJobs request.

ListDataScanJobsResponse

List DataScanJobs response.

ListDataScansRequest

List dataScans request.

ListDataScansResponse

List dataScans response.

ListDataTaxonomiesRequest

List DataTaxonomies request.

ListDataTaxonomiesResponse

List DataTaxonomies response.

ListEntitiesRequest

List metadata entities request.

ListEntitiesResponse

List metadata entities response.

ListEntriesRequest

ListEntriesResponse

ListEntryGroupsRequest

List entryGroups request.

ListEntryGroupsResponse

List ListEntryGroups response.

ListEntryTypesRequest

List EntryTypes request

ListEntryTypesResponse

List EntryTypes response

ListEnvironmentsRequest

List environments request.

ListEnvironmentsResponse

List environments response.

ListJobsRequest

List jobs request.

ListJobsResponse

List jobs response.

ListLakeActionsRequest

List lake actions request.

ListLakesRequest

List lakes request.

ListLakesResponse

List lakes response.

ListPartitionsRequest

List metadata partitions request.

ListPartitionsResponse

List metadata partitions response.

ListSessionsRequest

List sessions request.

ListSessionsResponse

List sessions response.

ListTasksRequest

List tasks request.

ListTasksResponse

List tasks response.

ListZoneActionsRequest

List zone actions request.

ListZonesRequest

List zones request.

ListZonesResponse

List zones response.

LookupEntryRequest

OperationMetadata

Represents the metadata of a long-running operation.

Partition

Represents partition metadata contained within entity instances.

ResourceAccessSpec

ResourceAccessSpec holds the access control configuration to be enforced on the resources, for example, Cloud Storage bucket, BigQuery dataset, BigQuery table.

RunDataScanRequest

Run DataScan Request

RunDataScanResponse

Run DataScan Response.

RunTaskRequest

RunTaskResponse

ScannedData

The data scanned during processing (e.g. in incremental DataScan)

Schema

Schema information describing the structure and layout of the data.

SearchEntriesRequest

SearchEntriesResponse

SearchEntriesResult

A single result of a SearchEntries request.

Session

Represents an active analyze session running for a user.

SessionEvent

These messages contain information about sessions within an environment. The monitored resource is ‘Environment’.

StorageAccess

Describes the access mechanism of the data within its storage location.

StorageFormat

Describes the format of the data within its storage location.

Task

A task represents a user-visible job.

Trigger

DataScan scheduling and trigger settings.

UpdateAspectTypeRequest

Update AspectType Request

UpdateAssetRequest

Update asset request.

UpdateContentRequest

Update content request.

UpdateDataAttributeBindingRequest

Update DataAttributeBinding request.

UpdateDataAttributeRequest

Update DataAttribute request.

UpdateDataScanRequest

Update dataScan request.

UpdateDataTaxonomyRequest

Update DataTaxonomy request.

UpdateEntityRequest

Update a metadata entity request. The exiting entity will be fully replaced by the entity in the request. The entity ID is mutable. To modify the ID, use the current entity ID in the request URL and specify the new ID in the request body.

UpdateEntryGroupRequest

Update EntryGroup Request

UpdateEntryRequest

UpdateEntryTypeRequest

Update EntryType Request

UpdateEnvironmentRequest

Update environment request.

UpdateLakeRequest

Update lake request.

UpdateTaskRequest

Update task request.

UpdateZoneRequest

Update zone request.

Zone

A zone represents a logical group of related assets within a lake. A zone can be used to map to organizational structure or represent stages of data readiness from raw to curated. It provides managing behavior that is shared or inherited by all contained assets.

Enums

DataScanType

The type of DataScan.

EntryView

View for controlling which parts of an entry are to be returned.

State

State of a resource.

StorageSystem

Identifies the cloud system that manages the data storage.

TransferStatus

Denotes the transfer status of a resource. It is unspecified for resources created from Dataplex API.