Module google_api_proto::google::cloud::dataplex::v1::asset

Nested message and enum types in Asset.

Modules

discovery_spec

Nested message and enum types in DiscoverySpec.

discovery_status

Nested message and enum types in DiscoveryStatus.

resource_spec

Nested message and enum types in ResourceSpec.

resource_status

Nested message and enum types in ResourceStatus.

security_status

Nested message and enum types in SecurityStatus.

Structs

DiscoverySpec

Settings to manage the metadata discovery and publishing for an asset.

DiscoveryStatus

Status of discovery for an asset.

ResourceSpec

Identifies the cloud resource that is referenced by this asset.

ResourceStatus

Status of the resource referenced by an asset.

SecurityStatus

Security policy status of the asset. Data security policy, i.e., readers, writers & owners, should be specified in the lake/zone/asset IAM policy.