Struct google_api_proto::google::cloud::dataplex::v1::catalog_service_client::CatalogServiceClient

pub struct CatalogServiceClient<T> { /* private fields */ }

The primary resources offered by this service are EntryGroups, EntryTypes, AspectTypes, Entry and Aspect which collectively allow a data administrator to organize, manage, secure and catalog data across their organization located across cloud projects in a variety of storage systems including Cloud Storage and BigQuery.

Implementations

impl<T> CatalogServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> CatalogServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_entry_type( &mut self, request: impl IntoRequest<CreateEntryTypeRequest>, ) -> Result<Response<Operation>, Status>

Creates an EntryType

pub async fn update_entry_type( &mut self, request: impl IntoRequest<UpdateEntryTypeRequest>, ) -> Result<Response<Operation>, Status>

Updates a EntryType resource.

pub async fn delete_entry_type( &mut self, request: impl IntoRequest<DeleteEntryTypeRequest>, ) -> Result<Response<Operation>, Status>

Deletes a EntryType resource.

pub async fn list_entry_types( &mut self, request: impl IntoRequest<ListEntryTypesRequest>, ) -> Result<Response<ListEntryTypesResponse>, Status>

Lists EntryType resources in a project and location.

pub async fn get_entry_type( &mut self, request: impl IntoRequest<GetEntryTypeRequest>, ) -> Result<Response<EntryType>, Status>

Retrieves a EntryType resource.

pub async fn create_aspect_type( &mut self, request: impl IntoRequest<CreateAspectTypeRequest>, ) -> Result<Response<Operation>, Status>

Creates an AspectType

pub async fn update_aspect_type( &mut self, request: impl IntoRequest<UpdateAspectTypeRequest>, ) -> Result<Response<Operation>, Status>

Updates a AspectType resource.

pub async fn delete_aspect_type( &mut self, request: impl IntoRequest<DeleteAspectTypeRequest>, ) -> Result<Response<Operation>, Status>

Deletes a AspectType resource.

pub async fn list_aspect_types( &mut self, request: impl IntoRequest<ListAspectTypesRequest>, ) -> Result<Response<ListAspectTypesResponse>, Status>

Lists AspectType resources in a project and location.

pub async fn get_aspect_type( &mut self, request: impl IntoRequest<GetAspectTypeRequest>, ) -> Result<Response<AspectType>, Status>

Retrieves a AspectType resource.

pub async fn create_entry_group( &mut self, request: impl IntoRequest<CreateEntryGroupRequest>, ) -> Result<Response<Operation>, Status>

Creates an EntryGroup

pub async fn update_entry_group( &mut self, request: impl IntoRequest<UpdateEntryGroupRequest>, ) -> Result<Response<Operation>, Status>

Updates a EntryGroup resource.

pub async fn delete_entry_group( &mut self, request: impl IntoRequest<DeleteEntryGroupRequest>, ) -> Result<Response<Operation>, Status>

Deletes a EntryGroup resource.

pub async fn list_entry_groups( &mut self, request: impl IntoRequest<ListEntryGroupsRequest>, ) -> Result<Response<ListEntryGroupsResponse>, Status>

Lists EntryGroup resources in a project and location.

pub async fn get_entry_group( &mut self, request: impl IntoRequest<GetEntryGroupRequest>, ) -> Result<Response<EntryGroup>, Status>

Retrieves a EntryGroup resource.

pub async fn create_entry( &mut self, request: impl IntoRequest<CreateEntryRequest>, ) -> Result<Response<Entry>, Status>

Creates an Entry.

pub async fn update_entry( &mut self, request: impl IntoRequest<UpdateEntryRequest>, ) -> Result<Response<Entry>, Status>

Updates an Entry.

pub async fn delete_entry( &mut self, request: impl IntoRequest<DeleteEntryRequest>, ) -> Result<Response<Entry>, Status>

Deletes an Entry.

pub async fn list_entries( &mut self, request: impl IntoRequest<ListEntriesRequest>, ) -> Result<Response<ListEntriesResponse>, Status>

Lists entries within an entry group.

pub async fn get_entry( &mut self, request: impl IntoRequest<GetEntryRequest>, ) -> Result<Response<Entry>, Status>

Gets a single entry.

pub async fn lookup_entry( &mut self, request: impl IntoRequest<LookupEntryRequest>, ) -> Result<Response<Entry>, Status>

Looks up a single entry.

pub async fn search_entries( &mut self, request: impl IntoRequest<SearchEntriesRequest>, ) -> Result<Response<SearchEntriesResponse>, Status>

Searches for entries matching given query and scope.