Struct google_api_proto::google::cloud::dataplex::v1::dataplex_service_client::DataplexServiceClient

pub struct DataplexServiceClient<T> { /* private fields */ }

Dataplex service provides data lakes as a service. The primary resources offered by this service are Lakes, Zones and Assets which collectively allow a data administrator to organize, manage, secure and catalog data across their organization located across cloud projects in a variety of storage systems including Cloud Storage and BigQuery.

Implementations

impl<T> DataplexServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> DataplexServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_lake( &mut self, request: impl IntoRequest<CreateLakeRequest>, ) -> Result<Response<Operation>, Status>

Creates a lake resource.

pub async fn update_lake( &mut self, request: impl IntoRequest<UpdateLakeRequest>, ) -> Result<Response<Operation>, Status>

Updates a lake resource.

pub async fn delete_lake( &mut self, request: impl IntoRequest<DeleteLakeRequest>, ) -> Result<Response<Operation>, Status>

Deletes a lake resource. All zones within the lake must be deleted before the lake can be deleted.

pub async fn list_lakes( &mut self, request: impl IntoRequest<ListLakesRequest>, ) -> Result<Response<ListLakesResponse>, Status>

Lists lake resources in a project and location.

pub async fn get_lake( &mut self, request: impl IntoRequest<GetLakeRequest>, ) -> Result<Response<Lake>, Status>

Retrieves a lake resource.

pub async fn list_lake_actions( &mut self, request: impl IntoRequest<ListLakeActionsRequest>, ) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in a lake.

pub async fn create_zone( &mut self, request: impl IntoRequest<CreateZoneRequest>, ) -> Result<Response<Operation>, Status>

Creates a zone resource within a lake.

pub async fn update_zone( &mut self, request: impl IntoRequest<UpdateZoneRequest>, ) -> Result<Response<Operation>, Status>

Updates a zone resource.

pub async fn delete_zone( &mut self, request: impl IntoRequest<DeleteZoneRequest>, ) -> Result<Response<Operation>, Status>

Deletes a zone resource. All assets within a zone must be deleted before the zone can be deleted.

pub async fn list_zones( &mut self, request: impl IntoRequest<ListZonesRequest>, ) -> Result<Response<ListZonesResponse>, Status>

Lists zone resources in a lake.

pub async fn get_zone( &mut self, request: impl IntoRequest<GetZoneRequest>, ) -> Result<Response<Zone>, Status>

Retrieves a zone resource.

pub async fn list_zone_actions( &mut self, request: impl IntoRequest<ListZoneActionsRequest>, ) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in a zone.

pub async fn create_asset( &mut self, request: impl IntoRequest<CreateAssetRequest>, ) -> Result<Response<Operation>, Status>

Creates an asset resource.

pub async fn update_asset( &mut self, request: impl IntoRequest<UpdateAssetRequest>, ) -> Result<Response<Operation>, Status>

Updates an asset resource.

pub async fn delete_asset( &mut self, request: impl IntoRequest<DeleteAssetRequest>, ) -> Result<Response<Operation>, Status>

Deletes an asset resource. The referenced storage resource is detached (default) or deleted based on the associated Lifecycle policy.

pub async fn list_assets( &mut self, request: impl IntoRequest<ListAssetsRequest>, ) -> Result<Response<ListAssetsResponse>, Status>

Lists asset resources in a zone.

pub async fn get_asset( &mut self, request: impl IntoRequest<GetAssetRequest>, ) -> Result<Response<Asset>, Status>

Retrieves an asset resource.

pub async fn list_asset_actions( &mut self, request: impl IntoRequest<ListAssetActionsRequest>, ) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in an asset.

pub async fn create_task( &mut self, request: impl IntoRequest<CreateTaskRequest>, ) -> Result<Response<Operation>, Status>

Creates a task resource within a lake.

pub async fn update_task( &mut self, request: impl IntoRequest<UpdateTaskRequest>, ) -> Result<Response<Operation>, Status>

Update the task resource.

pub async fn delete_task( &mut self, request: impl IntoRequest<DeleteTaskRequest>, ) -> Result<Response<Operation>, Status>

Delete the task resource.

pub async fn list_tasks( &mut self, request: impl IntoRequest<ListTasksRequest>, ) -> Result<Response<ListTasksResponse>, Status>

Lists tasks under the given lake.

pub async fn get_task( &mut self, request: impl IntoRequest<GetTaskRequest>, ) -> Result<Response<Task>, Status>

Get task resource.

pub async fn list_jobs( &mut self, request: impl IntoRequest<ListJobsRequest>, ) -> Result<Response<ListJobsResponse>, Status>

Lists Jobs under the given task.

pub async fn run_task( &mut self, request: impl IntoRequest<RunTaskRequest>, ) -> Result<Response<RunTaskResponse>, Status>

Run an on demand execution of a Task.

pub async fn get_job( &mut self, request: impl IntoRequest<GetJobRequest>, ) -> Result<Response<Job>, Status>

Get job resource.

pub async fn cancel_job( &mut self, request: impl IntoRequest<CancelJobRequest>, ) -> Result<Response<()>, Status>

Cancel jobs running for the task resource.

pub async fn create_environment( &mut self, request: impl IntoRequest<CreateEnvironmentRequest>, ) -> Result<Response<Operation>, Status>

Create an environment resource.

pub async fn update_environment( &mut self, request: impl IntoRequest<UpdateEnvironmentRequest>, ) -> Result<Response<Operation>, Status>

Update the environment resource.

pub async fn delete_environment( &mut self, request: impl IntoRequest<DeleteEnvironmentRequest>, ) -> Result<Response<Operation>, Status>

Delete the environment resource. All the child resources must have been deleted before environment deletion can be initiated.

pub async fn list_environments( &mut self, request: impl IntoRequest<ListEnvironmentsRequest>, ) -> Result<Response<ListEnvironmentsResponse>, Status>

Lists environments under the given lake.

pub async fn get_environment( &mut self, request: impl IntoRequest<GetEnvironmentRequest>, ) -> Result<Response<Environment>, Status>

Get environment resource.

pub async fn list_sessions( &mut self, request: impl IntoRequest<ListSessionsRequest>, ) -> Result<Response<ListSessionsResponse>, Status>

Lists session resources in an environment.