Module google_api_proto::google::cloud::dataproc::logging

Modules

autoscaler_recommendation

Nested message and enum types in AutoscalerRecommendation.

reconciliation_cluster_heal_log

Nested message and enum types in ReconciliationClusterHealLog.

reconciliation_log

Nested message and enum types in ReconciliationLog.

Structs

AutoscalerLog

The main proto that will be converted to JSON format and then written to Logging.

AutoscalerRecommendation

The inputs, outputs, and detailed explanation of the Autoscaling recommendation.

AutoscalerStatus

The Autoscaler’s status, including its state and other details.

ClusterSize

The short version of cluster configuration for Cloud Logging.

ReconciliationClusterHealLog

Reconciliation log for cluster heal event.

ReconciliationLog

Reconciliation log for session ttl event.

Enums

AutoscalerState

The Autoscaler state.

ConstrainingFactor

MetricType

The kind of metric input to the Autoscaling algorithm.

ReconciliationDecisionType

Decision type

ScalingDecisionType

The Autoscaling decision type.