Struct google_api_proto::google::cloud::dataproc::v1::ClusterConfig

pub struct ClusterConfig {

    pub config_bucket: String,
    pub temp_bucket: String,
    pub gce_cluster_config: Option<GceClusterConfig>,
    pub master_config: Option<InstanceGroupConfig>,
    pub worker_config: Option<InstanceGroupConfig>,
    pub secondary_worker_config: Option<InstanceGroupConfig>,
    pub software_config: Option<SoftwareConfig>,
    pub initialization_actions: Vec<NodeInitializationAction>,
    pub encryption_config: Option<EncryptionConfig>,
    pub autoscaling_config: Option<AutoscalingConfig>,
    pub security_config: Option<SecurityConfig>,
    pub lifecycle_config: Option<LifecycleConfig>,
    pub endpoint_config: Option<EndpointConfig>,
    pub metastore_config: Option<MetastoreConfig>,
    pub dataproc_metric_config: Option<DataprocMetricConfig>,
    pub auxiliary_node_groups: Vec<AuxiliaryNodeGroup>,

}

The cluster config.

Fields
config_bucket: String

Optional. A Cloud Storage bucket used to stage job dependencies, config files, and job driver console output. If you do not specify a staging bucket, Cloud Dataproc will determine a Cloud Storage location (US, ASIA, or EU) for your cluster’s staging bucket according to the Compute Engine zone where your cluster is deployed, and then create and manage this project-level, per-location bucket (see Dataproc staging and temp buckets). This field requires a Cloud Storage bucket name, not a gs://... URI to a Cloud Storage bucket.

temp_bucket: String

Optional. A Cloud Storage bucket used to store ephemeral cluster and jobs data, such as Spark and MapReduce history files. If you do not specify a temp bucket, Dataproc will determine a Cloud Storage location (US, ASIA, or EU) for your cluster’s temp bucket according to the Compute Engine zone where your cluster is deployed, and then create and manage this project-level, per-location bucket. The default bucket has a TTL of 90 days, but you can use any TTL (or none) if you specify a bucket (see Dataproc staging and temp buckets). This field requires a Cloud Storage bucket name, not a gs://... URI to a Cloud Storage bucket.

gce_cluster_config: Option<GceClusterConfig>

Optional. The shared Compute Engine config settings for all instances in a cluster.

master_config: Option<InstanceGroupConfig>

Optional. The Compute Engine config settings for the cluster’s master instance.

worker_config: Option<InstanceGroupConfig>

Optional. The Compute Engine config settings for the cluster’s worker instances.

secondary_worker_config: Option<InstanceGroupConfig>

Optional. The Compute Engine config settings for a cluster’s secondary worker instances

software_config: Option<SoftwareConfig>

Optional. The config settings for cluster software.

initialization_actions: Vec<NodeInitializationAction>

Optional. Commands to execute on each node after config is completed. By default, executables are run on master and all worker nodes. You can test a node’s role metadata to run an executable on a master or worker node, as shown below using curl (you can also use wget):

 ROLE=$(curl -H Metadata-Flavor:Google
 <http://metadata/computeMetadata/v1/instance/attributes/dataproc-role>)
 if \[[ "${ROLE}" == 'Master' ]\]; then
   ... master specific actions ...
 else
   ... worker specific actions ...
 fi

encryption_config: Option<EncryptionConfig>

Optional. Encryption settings for the cluster.

autoscaling_config: Option<AutoscalingConfig>

Optional. Autoscaling config for the policy associated with the cluster. Cluster does not autoscale if this field is unset.

security_config: Option<SecurityConfig>

Optional. Security settings for the cluster.

lifecycle_config: Option<LifecycleConfig>

Optional. Lifecycle setting for the cluster.

endpoint_config: Option<EndpointConfig>

Optional. Port/endpoint configuration for this cluster

metastore_config: Option<MetastoreConfig>

Optional. Metastore configuration.

dataproc_metric_config: Option<DataprocMetricConfig>

Optional. The config for Dataproc metrics.

auxiliary_node_groups: Vec<AuxiliaryNodeGroup>

Optional. The node group settings.