Struct google_api_proto::google::cloud::deploy::v1::cloud_deploy_client::CloudDeployClient

pub struct CloudDeployClient<T> { /* private fields */ }

CloudDeploy service creates and manages Continuous Delivery operations on Google Cloud Platform via Skaffold (https://skaffold.dev).

Implementations

impl<T> CloudDeployClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> CloudDeployClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_delivery_pipelines( &mut self, request: impl IntoRequest<ListDeliveryPipelinesRequest> ) -> Result<Response<ListDeliveryPipelinesResponse>, Status>

Lists DeliveryPipelines in a given project and location.

pub async fn get_delivery_pipeline( &mut self, request: impl IntoRequest<GetDeliveryPipelineRequest> ) -> Result<Response<DeliveryPipeline>, Status>

Gets details of a single DeliveryPipeline.

pub async fn create_delivery_pipeline( &mut self, request: impl IntoRequest<CreateDeliveryPipelineRequest> ) -> Result<Response<Operation>, Status>

Creates a new DeliveryPipeline in a given project and location.

pub async fn update_delivery_pipeline( &mut self, request: impl IntoRequest<UpdateDeliveryPipelineRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single DeliveryPipeline.

pub async fn delete_delivery_pipeline( &mut self, request: impl IntoRequest<DeleteDeliveryPipelineRequest> ) -> Result<Response<Operation>, Status>

Deletes a single DeliveryPipeline.

pub async fn list_targets( &mut self, request: impl IntoRequest<ListTargetsRequest> ) -> Result<Response<ListTargetsResponse>, Status>

Lists Targets in a given project and location.

pub async fn rollback_target( &mut self, request: impl IntoRequest<RollbackTargetRequest> ) -> Result<Response<RollbackTargetResponse>, Status>

Creates a Rollout to roll back the specified target.

pub async fn get_target( &mut self, request: impl IntoRequest<GetTargetRequest> ) -> Result<Response<Target>, Status>

Gets details of a single Target.

pub async fn create_target( &mut self, request: impl IntoRequest<CreateTargetRequest> ) -> Result<Response<Operation>, Status>

Creates a new Target in a given project and location.

pub async fn update_target( &mut self, request: impl IntoRequest<UpdateTargetRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Target.

pub async fn delete_target( &mut self, request: impl IntoRequest<DeleteTargetRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Target.

pub async fn list_custom_target_types( &mut self, request: impl IntoRequest<ListCustomTargetTypesRequest> ) -> Result<Response<ListCustomTargetTypesResponse>, Status>

Lists CustomTargetTypes in a given project and location.

pub async fn get_custom_target_type( &mut self, request: impl IntoRequest<GetCustomTargetTypeRequest> ) -> Result<Response<CustomTargetType>, Status>

Gets details of a single CustomTargetType.

pub async fn create_custom_target_type( &mut self, request: impl IntoRequest<CreateCustomTargetTypeRequest> ) -> Result<Response<Operation>, Status>

Creates a new CustomTargetType in a given project and location.

pub async fn update_custom_target_type( &mut self, request: impl IntoRequest<UpdateCustomTargetTypeRequest> ) -> Result<Response<Operation>, Status>

Updates a single CustomTargetType.

pub async fn delete_custom_target_type( &mut self, request: impl IntoRequest<DeleteCustomTargetTypeRequest> ) -> Result<Response<Operation>, Status>

Deletes a single CustomTargetType.

pub async fn list_releases( &mut self, request: impl IntoRequest<ListReleasesRequest> ) -> Result<Response<ListReleasesResponse>, Status>

Lists Releases in a given project and location.

pub async fn get_release( &mut self, request: impl IntoRequest<GetReleaseRequest> ) -> Result<Response<Release>, Status>

Gets details of a single Release.

pub async fn create_release( &mut self, request: impl IntoRequest<CreateReleaseRequest> ) -> Result<Response<Operation>, Status>

Creates a new Release in a given project and location.

pub async fn abandon_release( &mut self, request: impl IntoRequest<AbandonReleaseRequest> ) -> Result<Response<AbandonReleaseResponse>, Status>

Abandons a Release in the Delivery Pipeline.

pub async fn approve_rollout( &mut self, request: impl IntoRequest<ApproveRolloutRequest> ) -> Result<Response<ApproveRolloutResponse>, Status>

Approves a Rollout.

pub async fn advance_rollout( &mut self, request: impl IntoRequest<AdvanceRolloutRequest> ) -> Result<Response<AdvanceRolloutResponse>, Status>

Advances a Rollout in a given project and location.

pub async fn cancel_rollout( &mut self, request: impl IntoRequest<CancelRolloutRequest> ) -> Result<Response<CancelRolloutResponse>, Status>

Cancels a Rollout in a given project and location.

pub async fn list_rollouts( &mut self, request: impl IntoRequest<ListRolloutsRequest> ) -> Result<Response<ListRolloutsResponse>, Status>

Lists Rollouts in a given project and location.

pub async fn get_rollout( &mut self, request: impl IntoRequest<GetRolloutRequest> ) -> Result<Response<Rollout>, Status>

Gets details of a single Rollout.

pub async fn create_rollout( &mut self, request: impl IntoRequest<CreateRolloutRequest> ) -> Result<Response<Operation>, Status>

Creates a new Rollout in a given project and location.

pub async fn ignore_job( &mut self, request: impl IntoRequest<IgnoreJobRequest> ) -> Result<Response<IgnoreJobResponse>, Status>

Ignores the specified Job in a Rollout.

pub async fn retry_job( &mut self, request: impl IntoRequest<RetryJobRequest> ) -> Result<Response<RetryJobResponse>, Status>

Retries the specified Job in a Rollout.

pub async fn list_job_runs( &mut self, request: impl IntoRequest<ListJobRunsRequest> ) -> Result<Response<ListJobRunsResponse>, Status>

Lists JobRuns in a given project and location.

pub async fn get_job_run( &mut self, request: impl IntoRequest<GetJobRunRequest> ) -> Result<Response<JobRun>, Status>

Gets details of a single JobRun.

pub async fn terminate_job_run( &mut self, request: impl IntoRequest<TerminateJobRunRequest> ) -> Result<Response<TerminateJobRunResponse>, Status>

Terminates a Job Run in a given project and location.

pub async fn get_config( &mut self, request: impl IntoRequest<GetConfigRequest> ) -> Result<Response<Config>, Status>

Gets the configuration for a location.

pub async fn create_automation( &mut self, request: impl IntoRequest<CreateAutomationRequest> ) -> Result<Response<Operation>, Status>

Creates a new Automation in a given project and location.

pub async fn update_automation( &mut self, request: impl IntoRequest<UpdateAutomationRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Automation resource.

pub async fn delete_automation( &mut self, request: impl IntoRequest<DeleteAutomationRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Automation resource.

pub async fn get_automation( &mut self, request: impl IntoRequest<GetAutomationRequest> ) -> Result<Response<Automation>, Status>

Gets details of a single Automation.

pub async fn list_automations( &mut self, request: impl IntoRequest<ListAutomationsRequest> ) -> Result<Response<ListAutomationsResponse>, Status>

Lists Automations in a given project and location.

pub async fn get_automation_run( &mut self, request: impl IntoRequest<GetAutomationRunRequest> ) -> Result<Response<AutomationRun>, Status>

Gets details of a single AutomationRun.

pub async fn list_automation_runs( &mut self, request: impl IntoRequest<ListAutomationRunsRequest> ) -> Result<Response<ListAutomationRunsResponse>, Status>

Lists AutomationRuns in a given project and location.

pub async fn cancel_automation_run( &mut self, request: impl IntoRequest<CancelAutomationRunRequest> ) -> Result<Response<CancelAutomationRunResponse>, Status>

Cancels an AutomationRun. The state of the AutomationRun after cancelling is CANCELLED. CancelAutomationRun can be called on AutomationRun in the state IN_PROGRESS and PENDING; AutomationRun in a different state returns an FAILED_PRECONDITION error.