Module google_api_proto::google::cloud::developerconnect::v1

Modules

connection

Nested message and enum types in Connection.

developer_connect_client

Generated client implementations.

fetch_git_hub_installations_response

Nested message and enum types in FetchGitHubInstallationsResponse.

fetch_git_refs_request

Nested message and enum types in FetchGitRefsRequest.

git_hub_config

Nested message and enum types in GitHubConfig.

installation_state

Nested message and enum types in InstallationState.

Structs

Connection

Message describing Connection object

CreateConnectionRequest

Message for creating a Connection

CreateGitRepositoryLinkRequest

Message for creating a GitRepositoryLink

DeleteConnectionRequest

Message for deleting a Connection

DeleteGitRepositoryLinkRequest

Message for deleting a GitRepositoryLink

FetchGitHubInstallationsRequest

Request for fetching github installations.

FetchGitHubInstallationsResponse

Response of fetching github installations.

FetchGitRefsRequest

Request for fetching git refs.

FetchGitRefsResponse

Response for fetching git refs.

FetchLinkableGitRepositoriesRequest

Request message for FetchLinkableGitRepositoriesRequest.

FetchLinkableGitRepositoriesResponse

Response message for FetchLinkableGitRepositories.

FetchReadTokenRequest

Message for fetching SCM read token.

FetchReadTokenResponse

Message for responding to get read token.

FetchReadWriteTokenRequest

Message for fetching SCM read/write token.

FetchReadWriteTokenResponse

Message for responding to get read/write token.

GetConnectionRequest

Message for getting a Connection

GetGitRepositoryLinkRequest

Message for getting a GitRepositoryLink

GitHubConfig

Configuration for connections to github.com.

GitRepositoryLink

Message describing the GitRepositoryLink object

InstallationState

Describes stage and necessary actions to be taken by the user to complete the installation. Used for GitHub and GitHub Enterprise based connections.

LinkableGitRepository

LinkableGitRepository represents a git repository that can be linked to a connection.

ListConnectionsRequest

Message for requesting list of Connections

ListConnectionsResponse

Message for response to listing Connections

ListGitRepositoryLinksRequest

Message for requesting a list of GitRepositoryLinks

ListGitRepositoryLinksResponse

Message for response to listing GitRepositoryLinks

OAuthCredential

Represents an OAuth token of the account that authorized the Connection, and associated metadata.

OperationMetadata

Represents the metadata of the long-running operation.

UpdateConnectionRequest

Message for updating a Connection