Module google_api_proto::google::cloud::dialogflow::cx::v3::response_message

Nested message and enum types in ResponseMessage.

Modules

mixed_audio

Nested message and enum types in MixedAudio.

output_audio_text

Nested message and enum types in OutputAudioText.

telephony_transfer_call

Nested message and enum types in TelephonyTransferCall.

Structs

ConversationSuccess

Indicates that the conversation succeeded, i.e., the bot handled the issue that the customer talked to it about.

EndInteraction

Indicates that interaction with the Dialogflow agent has ended. This message is generated by Dialogflow only and not supposed to be defined by the user.

LiveAgentHandoff

Indicates that the conversation should be handed off to a live agent.

MixedAudio

Represents an audio message that is composed of both segments synthesized from the Dialogflow agent prompts and ones hosted externally at the specified URIs. The external URIs are specified via [play_audio][google.cloud.dialogflow.cx.v3.ResponseMessage.play_audio]. This message is generated by Dialogflow only and not supposed to be defined by the user.

OutputAudioText

A text or ssml response that is preferentially used for TTS output audio synthesis, as described in the comment on the ResponseMessage message.

PlayAudio

Specifies an audio clip to be played by the client as part of the response.

TelephonyTransferCall

Represents the signal that telles the client to transfer the phone call connected to the agent to a third-party endpoint.

Text

The text response message.

Enums

Message

Required. The rich response message.