Struct google_api_proto::google::cloud::dialogflow::cx::v3::QueryParameters

pub struct QueryParameters {

    pub time_zone: String,
    pub geo_location: Option<LatLng>,
    pub session_entity_types: Vec<SessionEntityType>,
    pub payload: Option<Struct>,
    pub parameters: Option<Struct>,
    pub current_page: String,
    pub disable_webhook: bool,
    pub analyze_query_text_sentiment: bool,
    pub webhook_headers: BTreeMap<String, String>,
    pub flow_versions: Vec<String>,
    pub channel: String,
    pub session_ttl: Option<Duration>,
    pub end_user_metadata: Option<Struct>,
    pub search_config: Option<SearchConfig>,

}

Represents the parameters of a conversational query.

Fields
time_zone: String

The time zone of this conversational query from the time zone database, e.g., America/New_York, Europe/Paris. If not provided, the time zone specified in the agent is used.

geo_location: Option<LatLng>

The geo location of this conversational query.

session_entity_types: Vec<SessionEntityType>

Additional session entity types to replace or extend developer entity types with. The entity synonyms apply to all languages and persist for the session of this query.

payload: Option<Struct>

This field can be used to pass custom data into the webhook associated with the agent. Arbitrary JSON objects are supported. Some integrations that query a Dialogflow agent may provide additional information in the payload. In particular, for the Dialogflow Phone Gateway integration, this field has the form:

{
  "telephony": {
    "caller_id": "+18558363987"
  }
}

parameters: Option<Struct>

Additional parameters to be put into [session parameters][SessionInfo.parameters]. To remove a parameter from the session, clients should explicitly set the parameter value to null.

You can reference the session parameters in the agent with the following format: $session.params.parameter-id.

Depending on your protocol or client library language, this is a map, associative array, symbol table, dictionary, or JSON object composed of a collection of (MapKey, MapValue) pairs:

MapKey type: string

MapKey value: parameter name

MapValue type: If parameter’s entity type is a composite entity then use map, otherwise, depending on the parameter value type, it could be one of string, number, boolean, null, list or map.

MapValue value: If parameter’s entity type is a composite entity then use map from composite entity property names to property values, otherwise, use parameter value.

current_page: String

The unique identifier of the [page][google.cloud.dialogflow.cx.v3.Page] to override the [current page][QueryResult.current_page] in the session. Format: projects/<Project ID>/locations/<Location ID>/agents/<Agent ID>/flows/<Flow ID>/pages/<Page ID>.

If current_page is specified, the previous state of the session will be ignored by Dialogflow, including the [previous page][QueryResult.current_page] and the [previous session parameters][QueryResult.parameters]. In most cases, [current_page][google.cloud.dialogflow.cx.v3.QueryParameters.current_page] and [parameters][google.cloud.dialogflow.cx.v3.QueryParameters.parameters] should be configured together to direct a session to a specific state.

disable_webhook: bool

Whether to disable webhook calls for this request.

analyze_query_text_sentiment: bool

Configures whether sentiment analysis should be performed. If not provided, sentiment analysis is not performed.

webhook_headers: BTreeMap<String, String>

This field can be used to pass HTTP headers for a webhook call. These headers will be sent to webhook along with the headers that have been configured through Dialogflow web console. The headers defined within this field will overwrite the headers configured through Dialogflow console if there is a conflict. Header names are case-insensitive. Google’s specified headers are not allowed. Including: “Host”, “Content-Length”, “Connection”, “From”, “User-Agent”, “Accept-Encoding”, “If-Modified-Since”, “If-None-Match”, “X-Forwarded-For”, etc.

flow_versions: Vec<String>

A list of flow versions to override for the request. Format: projects/<Project ID>/locations/<Location ID>/agents/<Agent ID>/flows/<Flow ID>/versions/<Version ID>.

If version 1 of flow X is included in this list, the traffic of flow X will go through version 1 regardless of the version configuration in the environment. Each flow can have at most one version specified in this list.

channel: String

The channel which this query is for.

If specified, only the [ResponseMessage][google.cloud.dialogflow.cx.v3.ResponseMessage] associated with the channel will be returned. If no [ResponseMessage][google.cloud.dialogflow.cx.v3.ResponseMessage] is associated with the channel, it falls back to the [ResponseMessage][google.cloud.dialogflow.cx.v3.ResponseMessage] with unspecified channel.

If unspecified, the [ResponseMessage][google.cloud.dialogflow.cx.v3.ResponseMessage] with unspecified channel will be returned.

session_ttl: Option<Duration>

Optional. Configure lifetime of the Dialogflow session. By default, a Dialogflow session remains active and its data is stored for 30 minutes after the last request is sent for the session. This value should be no longer than 1 day.

end_user_metadata: Option<Struct>

Optional. Information about the end-user to improve the relevance and accuracy of generative answers.

This will be interpreted and used by a language model, so, for good results, the data should be self-descriptive, and in a simple structure.

Example:

{
   "subscription plan": "Business Premium Plus",
   "devices owned": [
     {"model": "Google Pixel 7"},
     {"model": "Google Pixel Tablet"}
   ]
}

search_config: Option<SearchConfig>

Optional. Search configuration for UCS search queries.