Struct google_api_proto::google::cloud::dialogflow::cx::v3::QueryResult

pub struct QueryResult {

    pub language_code: String,
    pub parameters: Option<Struct>,
    pub response_messages: Vec<ResponseMessage>,
    pub webhook_statuses: Vec<Status>,
    pub webhook_payloads: Vec<Struct>,
    pub current_page: Option<Page>,
    pub intent: Option<Intent>,
    pub intent_detection_confidence: f32,
    pub match: Option<Match>,
    pub diagnostic_info: Option<Struct>,
    pub sentiment_analysis_result: Option<SentimentAnalysisResult>,
    pub advanced_settings: Option<AdvancedSettings>,
    pub allow_answer_feedback: bool,
    pub query: Option<Query>,

}

Represents the result of a conversational query.

Fields
language_code: String

The language that was triggered during intent detection. See Language Support for a list of the currently supported language codes.

parameters: Option<Struct>

The collected [session parameters][google.cloud.dialogflow.cx.v3.SessionInfo.parameters].

Depending on your protocol or client library language, this is a map, associative array, symbol table, dictionary, or JSON object composed of a collection of (MapKey, MapValue) pairs:

MapKey type: string

MapKey value: parameter name

MapValue type: If parameterโs entity type is a composite entity then use map, otherwise, depending on the parameter value type, it could be one of string, number, boolean, null, list or map.

MapValue value: If parameterโs entity type is a composite entity then use map from composite entity property names to property values, otherwise, use parameter value.

response_messages: Vec<ResponseMessage>

The list of rich messages returned to the client. Responses vary from simple text messages to more sophisticated, structured payloads used to drive complex logic.

webhook_statuses: Vec<Status>

The list of webhook call status in the order of call sequence.

webhook_payloads: Vec<Struct>

The list of webhook payload in [WebhookResponse.payload][google.cloud.dialogflow.cx.v3.WebhookResponse.payload], in the order of call sequence. If some webhook call fails or doesnโt return any payload, an empty Struct would be used instead.

current_page: Option<Page>

The current [Page][google.cloud.dialogflow.cx.v3.Page]. Some, not all fields are filled in this message, including but not limited to name and display_name.

intent: Option<Intent>
๐Deprecated

The [Intent][google.cloud.dialogflow.cx.v3.Intent] that matched the conversational query. Some, not all fields are filled in this message, including but not limited to: name and display_name. This field is deprecated, please use [QueryResult.match][google.cloud.dialogflow.cx.v3.QueryResult.match] instead.

intent_detection_confidence: f32
๐Deprecated

The intent detection confidence. Values range from 0.0 (completely uncertain) to 1.0 (completely certain). This value is for informational purpose only and is only used to help match the best intent within the classification threshold. This value may change for the same end-user expression at any time due to a model retraining or change in implementation. This field is deprecated, please use [QueryResult.match][google.cloud.dialogflow.cx.v3.QueryResult.match] instead.

match: Option<Match>

Intent match result, could be an intent or an event.

diagnostic_info: Option<Struct>

The free-form diagnostic info. For example, this field could contain webhook call latency. The fields of this data can change without notice, so you should not write code that depends on its structure.

One of the fields is called โAlternative Matched Intentsโ, which may aid with debugging. The following describes these intent results:

The list is empty if no intent was matched to end-user input.

Only intents that are referenced in the currently active flow are included.

The matched intent is included.

Other intents that could have matched end-user input, but did not match because they are referenced by intent routes that are out of scope, are included.

Other intents referenced by intent routes in scope that matched end-user input, but had a lower confidence score.

sentiment_analysis_result: Option<SentimentAnalysisResult>

The sentiment analyss result, which depends on [analyze_query_text_sentiment] [google.cloud.dialogflow.cx.v3.QueryParameters.analyze_query_text_sentiment], specified in the request.

advanced_settings: Option<AdvancedSettings>

Returns the current advanced settings including IVR settings. Even though the operations configured by these settings are performed by Dialogflow, the client may need to perform special logic at the moment. For example, if Dialogflow exports audio to Google Cloud Storage, then the client may need to wait for the resulting object to appear in the bucket before proceeding.

allow_answer_feedback: bool

Indicates whether the Thumbs up/Thumbs down rating controls are need to be shown for the response in the Dialogflow Messenger widget.

query: Option<Query>

The original conversational query.