Module google_api_proto::google::cloud::dialogflow::cx::v3beta1

Modules

action

Nested message and enum types in Action.

advanced_settings

Nested message and enum types in AdvancedSettings.

agent

Nested message and enum types in Agent.

agents_client

Generated client implementations.

answer_feedback

Nested message and enum types in AnswerFeedback.

boost_spec

Nested message and enum types in BoostSpec.

calculate_coverage_request

Nested message and enum types in CalculateCoverageRequest.

calculate_coverage_response

Nested message and enum types in CalculateCoverageResponse.

changelogs_client

Generated client implementations.

continuous_test_result

Nested message and enum types in ContinuousTestResult.

conversation

Nested message and enum types in Conversation.

conversation_history_client

Generated client implementations.

conversation_turn

Nested message and enum types in ConversationTurn.

data_store_connection_signals

Nested message and enum types in DataStoreConnectionSignals.

deployment

Nested message and enum types in Deployment.

deployments_client

Generated client implementations.

detect_intent_response

Nested message and enum types in DetectIntentResponse.

entity_type

Nested message and enum types in EntityType.

entity_types_client

Generated client implementations.

environment

Nested message and enum types in Environment.

environments_client

Generated client implementations.

event_handler

Nested message and enum types in EventHandler.

examples_client

Generated client implementations.

experiment

Nested message and enum types in Experiment.

experiments_client

Generated client implementations.

export_agent_request

Nested message and enum types in ExportAgentRequest.

export_agent_response

Nested message and enum types in ExportAgentResponse.

export_entity_types_request

Nested message and enum types in ExportEntityTypesRequest.

export_entity_types_response

Nested message and enum types in ExportEntityTypesResponse.

export_flow_response

Nested message and enum types in ExportFlowResponse.

export_intents_request

Nested message and enum types in ExportIntentsRequest.

export_intents_response

Nested message and enum types in ExportIntentsResponse.

export_test_cases_request

Nested message and enum types in ExportTestCasesRequest.

export_test_cases_response

Nested message and enum types in ExportTestCasesResponse.

export_tools_request

Nested message and enum types in ExportToolsRequest.

export_tools_response

Nested message and enum types in ExportToolsResponse.

flow

Nested message and enum types in Flow.

flows_client

Generated client implementations.

form

Nested message and enum types in Form.

fulfillment

Nested message and enum types in Fulfillment.

generative_settings

Nested message and enum types in GenerativeSettings.

generator

Nested message and enum types in Generator.

generators_client

Generated client implementations.

import_entity_types_request

Nested message and enum types in ImportEntityTypesRequest.

import_entity_types_response

Nested message and enum types in ImportEntityTypesResponse.

import_flow_request

Nested message and enum types in ImportFlowRequest.

import_intents_request

Nested message and enum types in ImportIntentsRequest.

import_intents_response

Nested message and enum types in ImportIntentsResponse.

import_test_cases_request

Nested message and enum types in ImportTestCasesRequest.

intent

Nested message and enum types in Intent.

intent_coverage

Nested message and enum types in IntentCoverage.

intents_client

Generated client implementations.

knowledge_connector_settings

Nested message and enum types in KnowledgeConnectorSettings.

list_test_cases_request

Nested message and enum types in ListTestCasesRequest.

match

Nested message and enum types in Match.

match_intent_response

Nested message and enum types in MatchIntentResponse.

nlu_settings

Nested message and enum types in NluSettings.

page_info

Nested message and enum types in PageInfo.

pages_client

Generated client implementations.

parameter_definition

Nested message and enum types in ParameterDefinition.

playbook

Nested message and enum types in Playbook.

playbooks_client

Generated client implementations.

query_input

Nested message and enum types in QueryInput.

query_result

Nested message and enum types in QueryResult.

response_message

Nested message and enum types in ResponseMessage.

restore_agent_request

Nested message and enum types in RestoreAgentRequest.

rollout_config

Nested message and enum types in RolloutConfig.

safety_settings

Nested message and enum types in SafetySettings.

security_settings

Nested message and enum types in SecuritySettings.

security_settings_service_client

Generated client implementations.

session_entity_type

Nested message and enum types in SessionEntityType.

session_entity_types_client

Generated client implementations.

sessions_client

Generated client implementations.

streaming_detect_intent_response

Nested message and enum types in StreamingDetectIntentResponse.

streaming_recognition_result

Nested message and enum types in StreamingRecognitionResult.

test_cases_client

Generated client implementations.

test_run_difference

Nested message and enum types in TestRunDifference.

tool

Nested message and enum types in Tool.

tool_call_result

Nested message and enum types in ToolCallResult.

tools_client

Generated client implementations.

transition_coverage

Nested message and enum types in TransitionCoverage.

transition_route

Nested message and enum types in TransitionRoute.

transition_route_group_coverage

Nested message and enum types in TransitionRouteGroupCoverage.

transition_route_groups_client

Generated client implementations.

validation_message

Nested message and enum types in ValidationMessage.

variants_history

Nested message and enum types in VariantsHistory.

version

Nested message and enum types in Version.

version_variants

Nested message and enum types in VersionVariants.

versions_client

Generated client implementations.

webhook

Nested message and enum types in Webhook.

webhook_request

Nested message and enum types in WebhookRequest.

webhook_response

Nested message and enum types in WebhookResponse.

webhooks_client

Generated client implementations.

Structs

Action

Action performed by end user or Dialogflow agent in the conversation.

AdvancedSettings

Hierarchical advanced settings for agent/flow/page/fulfillment/parameter. Settings exposed at lower level overrides the settings exposed at higher level. Overriding occurs at the sub-setting level. For example, the playback_interruption_settings at fulfillment level only overrides the playback_interruption_settings at the agent level, leaving other settings at the agent level unchanged.

Agent

Agents are best described as Natural Language Understanding (NLU) modules that transform user requests into actionable data. You can include agents in your app, product, or service to determine user intent and respond to the user in a natural way.

AgentUtterance

AgentUtterance represents one message sent by the agent.

AgentValidationResult

The response message for [Agents.GetAgentValidationResult][google.cloud.dialogflow.cx.v3beta1.Agents.GetAgentValidationResult].

AnswerFeedback

Stores information about feedback provided by users about a response.

AudioInput

Represents the natural speech audio to be processed.

BargeInConfig

Configuration of the barge-in behavior. Barge-in instructs the API to return a detected utterance at a proper time while the client is playing back the response audio from a previous request. When the client sees the utterance, it should stop the playback and immediately get ready for receiving the responses for the current request.

BatchDeleteTestCasesRequest

The request message for [TestCases.BatchDeleteTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.BatchDeleteTestCases].

BatchRunTestCasesMetadata

Metadata returned for the [TestCases.BatchRunTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.BatchRunTestCases] long running operation.

BatchRunTestCasesRequest

The request message for [TestCases.BatchRunTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.BatchRunTestCases].

BatchRunTestCasesResponse

The response message for [TestCases.BatchRunTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.BatchRunTestCases].

BoostSpec

Boost specification to boost certain documents. A copy of google.cloud.discoveryengine.v1main.BoostSpec, field documentation is available at https://cloud.google.com/generative-ai-app-builder/docs/reference/rest/v1alpha/BoostSpec

BoostSpecs

Boost specifications for data stores.

CalculateCoverageRequest

The request message for [TestCases.CalculateCoverage][google.cloud.dialogflow.cx.v3beta1.TestCases.CalculateCoverage].

CalculateCoverageResponse

The response message for [TestCases.CalculateCoverage][google.cloud.dialogflow.cx.v3beta1.TestCases.CalculateCoverage].

Changelog

Changelogs represents a change made to a given agent.

CloudConversationDebuggingInfo

Cloud conversation info for easier debugging. It will get populated in StreamingDetectIntentResponse or StreamingAnalyzeContentResponse when the flag enable_debugging_info is set to true in corresponding requests.

CompareVersionsRequest

The request message for [Versions.CompareVersions][google.cloud.dialogflow.cx.v3beta1.Versions.CompareVersions].

CompareVersionsResponse

The response message for [Versions.CompareVersions][google.cloud.dialogflow.cx.v3beta1.Versions.CompareVersions].

ContinuousTestResult

Represents a result from running a test case in an agent environment.

Conversation

Represents a conversation.

ConversationTurn

One interaction between a human and virtual agent. The human provides some input and the virtual agent provides a response.

CreateAgentRequest

The request message for [Agents.CreateAgent][google.cloud.dialogflow.cx.v3beta1.Agents.CreateAgent].

CreateEntityTypeRequest

The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.CreateEntityType].

CreateEnvironmentRequest

The request message for [Environments.CreateEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.CreateEnvironment].

CreateExampleRequest

The request message for [Examples.CreateExample][google.cloud.dialogflow.cx.v3beta1.Examples.CreateExample].

CreateExperimentRequest

The request message for [Experiments.CreateExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.CreateExperiment].

CreateFlowRequest

The request message for [Flows.CreateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.CreateFlow].

CreateGeneratorRequest

The request message for [Generators.CreateGenerator][google.cloud.dialogflow.cx.v3beta1.Generators.CreateGenerator].

CreateIntentRequest

The request message for [Intents.CreateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.CreateIntent].

CreatePageRequest

The request message for [Pages.CreatePage][google.cloud.dialogflow.cx.v3beta1.Pages.CreatePage].

CreatePlaybookRequest

The request message for [Playbooks.CreatePlaybook][google.cloud.dialogflow.cx.v3beta1.Playbooks.CreatePlaybook].

CreatePlaybookVersionRequest

The request message for [Playbooks.CreatePlaybookVersion][google.cloud.dialogflow.cx.v3beta1.Playbooks.CreatePlaybookVersion].

CreateSecuritySettingsRequest

The request message for [SecuritySettings.CreateSecuritySettings][].

CreateSessionEntityTypeRequest

The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.CreateSessionEntityType].

CreateTestCaseRequest

The request message for [TestCases.CreateTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.CreateTestCase].

CreateToolRequest

The request message for [Tools.CreateTool][google.cloud.dialogflow.cx.v3beta1.Tools.CreateTool].

CreateTransitionRouteGroupRequest

The request message for [TransitionRouteGroups.CreateTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.CreateTransitionRouteGroup].

CreateVersionOperationMetadata

Metadata associated with the long running operation for [Versions.CreateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.CreateVersion].

CreateVersionRequest

The request message for [Versions.CreateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.CreateVersion].

CreateWebhookRequest

The request message for [Webhooks.CreateWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.CreateWebhook].

DataStoreConnection

A data store connection. It represents a data store in Discovery Engine and the type of the contents it contains.

DataStoreConnectionSignals

Data store connection feature output signals. Might be only partially field if processing stop before the final answer. Reasons for this can be, but are not limited to: empty UCS search results, positive RAI check outcome, grounding failure, …

DeleteAgentRequest

The request message for [Agents.DeleteAgent][google.cloud.dialogflow.cx.v3beta1.Agents.DeleteAgent].

DeleteConversationRequest

The request message for [Conversations.DeleteConversation][].

DeleteEntityTypeRequest

The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.DeleteEntityType].

DeleteEnvironmentRequest

The request message for [Environments.DeleteEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.DeleteEnvironment].

DeleteExampleRequest

The request message for [Examples.DeleteExample][google.cloud.dialogflow.cx.v3beta1.Examples.DeleteExample].

DeleteExperimentRequest

The request message for [Experiments.DeleteExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.DeleteExperiment].

DeleteFlowRequest

The request message for [Flows.DeleteFlow][google.cloud.dialogflow.cx.v3beta1.Flows.DeleteFlow].

DeleteGeneratorRequest

The request message for [Generators.DeleteGenerator][google.cloud.dialogflow.cx.v3beta1.Generators.DeleteGenerator].

DeleteIntentRequest

The request message for [Intents.DeleteIntent][google.cloud.dialogflow.cx.v3beta1.Intents.DeleteIntent].

DeletePageRequest

The request message for [Pages.DeletePage][google.cloud.dialogflow.cx.v3beta1.Pages.DeletePage].

DeletePlaybookRequest

The request message for [Playbooks.DeletePlaybook][google.cloud.dialogflow.cx.v3beta1.Playbooks.DeletePlaybook].

DeletePlaybookVersionRequest

The request message for [Playbooks.DeletePlaybookVersion][google.cloud.dialogflow.cx.v3beta1.Playbooks.DeletePlaybookVersion].

DeleteSecuritySettingsRequest

The request message for [SecuritySettings.DeleteSecuritySettings][].

DeleteSessionEntityTypeRequest

The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.DeleteSessionEntityType].

DeleteToolRequest

The request message for [Tools.DeleteTool][google.cloud.dialogflow.cx.v3beta1.Tools.DeleteTool].

DeleteTransitionRouteGroupRequest

The request message for [TransitionRouteGroups.DeleteTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.DeleteTransitionRouteGroup].

DeleteVersionRequest

The request message for [Versions.DeleteVersion][google.cloud.dialogflow.cx.v3beta1.Versions.DeleteVersion].

DeleteWebhookRequest

The request message for [Webhooks.DeleteWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.DeleteWebhook].

DeployFlowMetadata

Metadata returned for the [Environments.DeployFlow][google.cloud.dialogflow.cx.v3beta1.Environments.DeployFlow] long running operation.

DeployFlowRequest

The request message for [Environments.DeployFlow][google.cloud.dialogflow.cx.v3beta1.Environments.DeployFlow].

DeployFlowResponse

The response message for [Environments.DeployFlow][google.cloud.dialogflow.cx.v3beta1.Environments.DeployFlow].

Deployment

Represents a deployment in an environment. A deployment happens when a flow version configured to be active in the environment. You can configure running pre-deployment steps, e.g. running validation test cases, experiment auto-rollout, etc.

DetectIntentRequest

The request to detect user’s intent.

DetectIntentResponse

The message returned from the DetectIntent method.

DtmfInput

Represents the input for dtmf event.

EntityType

Entities are extracted from user input and represent parameters that are meaningful to your application. For example, a date range, a proper name such as a geographic location or landmark, and so on. Entities represent actionable data for your application.

Environment

Represents an environment for an agent. You can create multiple versions of your agent and publish them to separate environments. When you edit an agent, you are editing the draft agent. At any point, you can save the draft agent as an agent version, which is an immutable snapshot of your agent. When you save the draft agent, it is published to the default environment. When you create agent versions, you can publish them to custom environments. You can create a variety of custom environments for testing, development, production, etc.

EventHandler

An event handler specifies an [event][google.cloud.dialogflow.cx.v3beta1.EventHandler.event] that can be handled during a session. When the specified event happens, the following actions are taken in order:

EventInput

Represents the event to trigger.

Example

Example represents a sample execution of the playbook in the conversation.

Experiment

Represents an experiment in an environment.

ExportAgentRequest

The request message for [Agents.ExportAgent][google.cloud.dialogflow.cx.v3beta1.Agents.ExportAgent].

ExportAgentResponse

The response message for [Agents.ExportAgent][google.cloud.dialogflow.cx.v3beta1.Agents.ExportAgent].

ExportEntityTypesMetadata

Metadata returned for the [EntityTypes.ExportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ExportEntityTypes] long running operation.

ExportEntityTypesRequest

The request message for [EntityTypes.ExportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ExportEntityTypes].

ExportEntityTypesResponse

The response message for [EntityTypes.ExportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ExportEntityTypes].

ExportFlowRequest

The request message for [Flows.ExportFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ExportFlow].

ExportFlowResponse

The response message for [Flows.ExportFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ExportFlow].

ExportIntentsMetadata

Metadata returned for the [Intents.ExportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ExportIntents] long running operation.

ExportIntentsRequest

The request message for [Intents.ExportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ExportIntents].

ExportIntentsResponse

The response message for [Intents.ExportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ExportIntents].

ExportTestCasesMetadata

Metadata returned for the [TestCases.ExportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ExportTestCases] long running operation. This message currently has no fields.

ExportTestCasesRequest

The request message for [TestCases.ExportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ExportTestCases].

ExportTestCasesResponse

The response message for [TestCases.ExportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ExportTestCases].

ExportToolsMetadata

Metadata returned for the [Tools.ExportTools][google.cloud.dialogflow.cx.v3beta1.Tools.ExportTools] long running operation.

ExportToolsRequest

The request message for [Tools.ExportTools][google.cloud.dialogflow.cx.v3beta1.Tools.ExportTools].

ExportToolsResponse

The response message for [Tools.ExportTools][google.cloud.dialogflow.cx.v3beta1.Tools.ExportTools].

FilterSpecs

Filter specifications for data stores.

Flow

Flows represents the conversation flows when you build your chatbot agent.

FlowImportStrategy

The flow import strategy used for resource conflict resolution associated with an [ImportFlowRequest][google.cloud.dialogflow.cx.v3beta1.ImportFlowRequest].

FlowInvocation

Stores metadata of the invocation of a CX flow. Next Id: 7

FlowValidationResult

The response message for [Flows.GetFlowValidationResult][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlowValidationResult].

Form

A form is a data model that groups related parameters that can be collected from the user. The process in which the agent prompts the user and collects parameter values from the user is called form filling. A form can be added to a [page][google.cloud.dialogflow.cx.v3beta1.Page]. When form filling is done, the filled parameters will be written to the [session][google.cloud.dialogflow.cx.v3beta1.SessionInfo.parameters].

FulfillIntentRequest

Request of [FulfillIntent][]

FulfillIntentResponse

Response of [FulfillIntent][]

Fulfillment

A fulfillment can do one or more of the following actions at the same time:

GcsDestination

Google Cloud Storage location for a Dialogflow operation that writes or exports objects (e.g. exported agent or transcripts) outside of Dialogflow.

GenerativeInfo

Represents the information of a query if handled by generative agent resources.

GenerativeSettings

Settings for Generative AI.

Generator

Generators contain prompt to be sent to the LLM model to generate text. The prompt can contain parameters which will be resolved before calling the model. It can optionally contain banned phrases to ensure the model responses are safe.

GetAgentRequest

The request message for [Agents.GetAgent][google.cloud.dialogflow.cx.v3beta1.Agents.GetAgent].

GetAgentValidationResultRequest

The request message for [Agents.GetAgentValidationResult][google.cloud.dialogflow.cx.v3beta1.Agents.GetAgentValidationResult].

GetChangelogRequest

The request message for [Changelogs.GetChangelog][google.cloud.dialogflow.cx.v3beta1.Changelogs.GetChangelog].

GetConversationRequest

The request message for [Conversations.GetConversation][].

GetDeploymentRequest

The request message for [Deployments.GetDeployment][google.cloud.dialogflow.cx.v3beta1.Deployments.GetDeployment].

GetEntityTypeRequest

The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.GetEntityType].

GetEnvironmentRequest

The request message for [Environments.GetEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.GetEnvironment].

GetExampleRequest

The request message for [Examples.GetExample][google.cloud.dialogflow.cx.v3beta1.Examples.GetExample].

GetExperimentRequest

The request message for [Experiments.GetExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.GetExperiment].

GetFlowRequest

The response message for [Flows.GetFlow][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlow].

GetFlowValidationResultRequest

The request message for [Flows.GetFlowValidationResult][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlowValidationResult].

GetGenerativeSettingsRequest

Request for [GetGenerativeSettings][google.cloud.dialogflow.cx.v3beta1.Agents.GetGenerativeSettings] RPC.

GetGeneratorRequest

The request message for [Generators.GetGenerator][google.cloud.dialogflow.cx.v3beta1.Generators.GetGenerator].

GetIntentRequest

The request message for [Intents.GetIntent][google.cloud.dialogflow.cx.v3beta1.Intents.GetIntent].

GetPageRequest

The request message for [Pages.GetPage][google.cloud.dialogflow.cx.v3beta1.Pages.GetPage].

GetPlaybookRequest

The request message for [Playbooks.GetPlaybook][google.cloud.dialogflow.cx.v3beta1.Playbooks.GetPlaybook].

GetPlaybookVersionRequest

The request message for [Playbooks.GetPlaybookVersion][google.cloud.dialogflow.cx.v3beta1.Playbooks.GetPlaybookVersion].

GetSecuritySettingsRequest

The request message for [SecuritySettingsService.GetSecuritySettings][google.cloud.dialogflow.cx.v3beta1.SecuritySettingsService.GetSecuritySettings].

GetSessionEntityTypeRequest

The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.GetSessionEntityType].

GetTestCaseRequest

The request message for [TestCases.GetTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.GetTestCase].

GetTestCaseResultRequest

The request message for [TestCases.GetTestCaseResult][google.cloud.dialogflow.cx.v3beta1.TestCases.GetTestCaseResult].

GetToolRequest

The request message for [Tools.GetTool][google.cloud.dialogflow.cx.v3beta1.Tools.GetTool].

GetTransitionRouteGroupRequest

The request message for [TransitionRouteGroups.GetTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.GetTransitionRouteGroup].

GetVersionRequest

The request message for [Versions.GetVersion][google.cloud.dialogflow.cx.v3beta1.Versions.GetVersion].

GetWebhookRequest

The request message for [Webhooks.GetWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.GetWebhook].

ImportEntityTypesMetadata

Metadata returned for the [EntityTypes.ImportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ImportEntityTypes] long running operation.

ImportEntityTypesRequest

The request message for [EntityTypes.ImportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ImportEntityTypes].

ImportEntityTypesResponse

The response message for [EntityTypes.ImportEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ImportEntityTypes].

ImportFlowRequest

The request message for [Flows.ImportFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ImportFlow].

ImportFlowResponse

The response message for [Flows.ImportFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ImportFlow].

ImportIntentsMetadata

Metadata returned for the [Intents.ImportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ImportIntents] long running operation.

ImportIntentsRequest

The request message for [Intents.ImportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ImportIntents].

ImportIntentsResponse

The response message for [Intents.ImportIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ImportIntents].

ImportTestCasesMetadata

Metadata returned for the [TestCases.ImportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ImportTestCases] long running operation.

ImportTestCasesRequest

The request message for [TestCases.ImportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ImportTestCases].

ImportTestCasesResponse

The response message for [TestCases.ImportTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ImportTestCases].

InlineDestination

Inline destination for a Dialogflow operation that writes or exports objects (e.g. [intents][google.cloud.dialogflow.cx.v3beta1.Intent]) outside of Dialogflow.

InlineSource

Inline source for a Dialogflow operation that reads or imports objects (e.g. [intents][google.cloud.dialogflow.cx.v3beta1.Intent]) into Dialogflow.

InputAudioConfig

Instructs the speech recognizer on how to process the audio content.

Intent

An intent represents a user’s intent to interact with a conversational agent.

IntentCoverage

Intent coverage represents the percentage of all possible intents in the agent that are triggered in any of a parent’s test cases.

IntentInput

Represents the intent to trigger programmatically rather than as a result of natural language processing.

KnowledgeConnectorSettings

The Knowledge Connector settings for this page or flow. This includes information such as the attached Knowledge Bases, and the way to execute fulfillment.

LanguageInfo

Represents the language information of the request.

ListAgentsRequest

The request message for [Agents.ListAgents][google.cloud.dialogflow.cx.v3beta1.Agents.ListAgents].

ListAgentsResponse

The response message for [Agents.ListAgents][google.cloud.dialogflow.cx.v3beta1.Agents.ListAgents].

ListChangelogsRequest

The request message for [Changelogs.ListChangelogs][google.cloud.dialogflow.cx.v3beta1.Changelogs.ListChangelogs].

ListChangelogsResponse

The response message for [Changelogs.ListChangelogs][google.cloud.dialogflow.cx.v3beta1.Changelogs.ListChangelogs].

ListContinuousTestResultsRequest

The request message for [Environments.ListContinuousTestResults][google.cloud.dialogflow.cx.v3beta1.Environments.ListContinuousTestResults].

ListContinuousTestResultsResponse

The response message for [Environments.ListTestCaseResults][].

ListConversationsRequest

The request message for [Conversations.ListConversations][].

ListConversationsResponse

The response message for [Conversations.ListConversations][].

ListDeploymentsRequest

The request message for [Deployments.ListDeployments][google.cloud.dialogflow.cx.v3beta1.Deployments.ListDeployments].

ListDeploymentsResponse

The response message for [Deployments.ListDeployments][google.cloud.dialogflow.cx.v3beta1.Deployments.ListDeployments].

ListEntityTypesRequest

The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListEntityTypesResponse

The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListEnvironmentsRequest

The request message for [Environments.ListEnvironments][google.cloud.dialogflow.cx.v3beta1.Environments.ListEnvironments].

ListEnvironmentsResponse

The response message for [Environments.ListEnvironments][google.cloud.dialogflow.cx.v3beta1.Environments.ListEnvironments].

ListExamplesRequest

The request message for [Examples.ListExamples][google.cloud.dialogflow.cx.v3beta1.Examples.ListExamples].

ListExamplesResponse

The response message for [Examples.ListExamples][google.cloud.dialogflow.cx.v3beta1.Examples.ListExamples].

ListExperimentsRequest

The request message for [Experiments.ListExperiments][google.cloud.dialogflow.cx.v3beta1.Experiments.ListExperiments].

ListExperimentsResponse

The response message for [Experiments.ListExperiments][google.cloud.dialogflow.cx.v3beta1.Experiments.ListExperiments].

ListFlowsRequest

The request message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListFlowsResponse

The response message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListGeneratorsRequest

The request message for [Generators.ListGenerators][google.cloud.dialogflow.cx.v3beta1.Generators.ListGenerators].

ListGeneratorsResponse

The response message for [Generators.ListGenerators][google.cloud.dialogflow.cx.v3beta1.Generators.ListGenerators].

ListIntentsRequest

The request message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListIntentsResponse

The response message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListPagesRequest

The request message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListPagesResponse

The response message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListPlaybookVersionsRequest

The request message for [Playbooks.ListPlaybookVersions][google.cloud.dialogflow.cx.v3beta1.Playbooks.ListPlaybookVersions].

ListPlaybookVersionsResponse

The response message for [Playbooks.ListPlaybookVersions][google.cloud.dialogflow.cx.v3beta1.Playbooks.ListPlaybookVersions].

ListPlaybooksRequest

The request message for [Playbooks.ListPlaybooks][google.cloud.dialogflow.cx.v3beta1.Playbooks.ListPlaybooks].

ListPlaybooksResponse

The response message for [Playbooks.ListPlaybooks][google.cloud.dialogflow.cx.v3beta1.Playbooks.ListPlaybooks].

ListSecuritySettingsRequest

The request message for [SecuritySettings.ListSecuritySettings][].

ListSecuritySettingsResponse

The response message for [SecuritySettings.ListSecuritySettings][].

ListSessionEntityTypesRequest

The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse

The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

ListTestCaseResultsRequest

The request message for [TestCases.ListTestCaseResults][google.cloud.dialogflow.cx.v3beta1.TestCases.ListTestCaseResults].

ListTestCaseResultsResponse

The response message for [TestCases.ListTestCaseResults][google.cloud.dialogflow.cx.v3beta1.TestCases.ListTestCaseResults].

ListTestCasesRequest

The request message for [TestCases.ListTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ListTestCases].

ListTestCasesResponse

The response message for [TestCases.ListTestCases][google.cloud.dialogflow.cx.v3beta1.TestCases.ListTestCases].

ListToolsRequest

The request message for [Tools.ListTools][google.cloud.dialogflow.cx.v3beta1.Tools.ListTools].

ListToolsResponse

The response message for [Tools.ListTools][google.cloud.dialogflow.cx.v3beta1.Tools.ListTools].

ListTransitionRouteGroupsRequest

The request message for [TransitionRouteGroups.ListTransitionRouteGroups][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.ListTransitionRouteGroups].

ListTransitionRouteGroupsResponse

The response message for [TransitionRouteGroups.ListTransitionRouteGroups][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.ListTransitionRouteGroups].

ListVersionsRequest

The request message for [Versions.ListVersions][google.cloud.dialogflow.cx.v3beta1.Versions.ListVersions].

ListVersionsResponse

The response message for [Versions.ListVersions][google.cloud.dialogflow.cx.v3beta1.Versions.ListVersions].

ListWebhooksRequest

The request message for [Webhooks.ListWebhooks][google.cloud.dialogflow.cx.v3beta1.Webhooks.ListWebhooks].

ListWebhooksResponse

The response message for [Webhooks.ListWebhooks][google.cloud.dialogflow.cx.v3beta1.Webhooks.ListWebhooks].

LlmModelSettings

Settings for LLM models.

LoadVersionRequest

The request message for [Versions.LoadVersion][google.cloud.dialogflow.cx.v3beta1.Versions.LoadVersion].

LookupEnvironmentHistoryRequest

The request message for [Environments.LookupEnvironmentHistory][google.cloud.dialogflow.cx.v3beta1.Environments.LookupEnvironmentHistory].

LookupEnvironmentHistoryResponse

The response message for [Environments.LookupEnvironmentHistory][google.cloud.dialogflow.cx.v3beta1.Environments.LookupEnvironmentHistory].

Match

Represents one match result of [MatchIntent][].

MatchIntentRequest

Request of [MatchIntent][].

MatchIntentResponse

Response of [MatchIntent][].

NluSettings

Settings related to NLU.

OutputAudioConfig

Instructs the speech synthesizer how to generate the output audio content.

Page

A Dialogflow CX conversation (session) can be described and visualized as a state machine. The states of a CX session are represented by pages.

PageInfo

Represents page information communicated to and from the webhook.

ParameterDefinition

Defines the properties of a parameter. Used to define parameters used in the agent and the input / output parameters for each fulfillment. (– Next Id: 4 –) (– api-linter: core::0123::resource-annotation=disabled aip.dev/not-precedent: ParameterDefinition is not an exposed resource. –)

Phrase

Text input which can be used for prompt or banned phrases.

Playbook

Playbook is the basic building block to instruct the LLM how to execute a certain task.

PlaybookInput

Input of the playbook.

PlaybookInvocation

Stores metadata of the invocation of a child playbook. Next Id: 5

PlaybookOutput

Output of the playbook.

PlaybookVersion

Playbook version is a snapshot of the playbook at certain timestamp.

QueryInput

Represents the query input. It can contain one of:

QueryParameters

Represents the parameters of a conversational query.

QueryResult

Represents the result of a conversational query.

ResourceName

Resource name and display name.

ResponseMessage

Represents a response message that can be returned by a conversational agent.

RestoreAgentRequest

The request message for [Agents.RestoreAgent][google.cloud.dialogflow.cx.v3beta1.Agents.RestoreAgent].

RolloutConfig

The configuration for auto rollout.

RolloutState

State of the auto-rollout process.

RunContinuousTestMetadata

Metadata returned for the [Environments.RunContinuousTest][google.cloud.dialogflow.cx.v3beta1.Environments.RunContinuousTest] long running operation.

RunContinuousTestRequest

The request message for [Environments.RunContinuousTest][google.cloud.dialogflow.cx.v3beta1.Environments.RunContinuousTest].

RunContinuousTestResponse

The response message for [Environments.RunContinuousTest][google.cloud.dialogflow.cx.v3beta1.Environments.RunContinuousTest].

RunTestCaseMetadata

Metadata returned for the [TestCases.RunTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.RunTestCase] long running operation. This message currently has no fields.

RunTestCaseRequest

The request message for [TestCases.RunTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.RunTestCase].

RunTestCaseResponse

The response message for [TestCases.RunTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.RunTestCase].

SafetySettings

Settings for Generative Safety.

SearchConfig

Search configuration for UCS search queries.

SecuritySettings

Represents the settings related to security issues, such as data redaction and data retention. It may take hours for updates on the settings to propagate to all the related components and take effect.

SentimentAnalysisResult

The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user’s attitude as positive, negative, or neutral.

SessionEntityType

Session entity types are referred to as User entity types and are entities that are built for an individual user such as favorites, preferences, playlists, and so on.

SessionInfo

Represents session information communicated to and from the webhook.

SpeechToTextSettings

Settings related to speech recognition.

SpeechWordInfo

Information for a word recognized by the speech recognizer.

StartExperimentRequest

The request message for [Experiments.StartExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.StartExperiment].

StopExperimentRequest

The request message for [Experiments.StopExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.StopExperiment].

StreamingDetectIntentRequest

The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.cx.v3beta1.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse

The top-level message returned from the [StreamingDetectIntent][google.cloud.dialogflow.cx.v3beta1.Sessions.StreamingDetectIntent] method.

StreamingRecognitionResult

Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SubmitAnswerFeedbackRequest

The request to set the feedback for a bot answer.

SynthesizeSpeechConfig

Configuration of how speech should be synthesized.

TestCase

Represents a test case.

TestCaseError

Error info for importing a test.

TestCaseResult

Represents a result from running a test case in an agent environment.

TestConfig

Represents configurations for a test case.

TestError

Error info for running a test.

TestRunDifference

The description of differences between original and replayed agent output.

TextInput

Represents the natural language text to be processed.

TextToSpeechSettings

Settings related to speech synthesizing.

Tool

A tool provides a list of actions which are available to the [Playbook][google.cloud.dialogflow.cx.v3beta1.Playbook] to attain its goal. A Tool consists of a description of the tool’s usage and a specification of the tool which contains the schema and authentication information.

ToolCall

Represents a call of a specific tool’s action with the specified inputs.

ToolCallResult

The result of calling a tool’s action that has been executed by the client.

ToolUse

Stores metadata of the invocation of an action supported by a tool.

TrainFlowRequest

The request message for [Flows.TrainFlow][google.cloud.dialogflow.cx.v3beta1.Flows.TrainFlow].

TransitionCoverage

Transition coverage represents the percentage of all possible page transitions (page-level transition routes and event handlers, excluding transition route groups) present within any of a parent’s test cases.

TransitionRoute

A transition route specifies a [intent][google.cloud.dialogflow.cx.v3beta1.Intent] that can be matched and/or a data condition that can be evaluated during a session. When a specified transition is matched, the following actions are taken in order:

TransitionRouteGroup

A TransitionRouteGroup represents a group of [TransitionRoutes][google.cloud.dialogflow.cx.v3beta1.TransitionRoute] to be used by a [Page][google.cloud.dialogflow.cx.v3beta1.Page].

TransitionRouteGroupCoverage

Transition route group coverage represents the percentage of all possible transition routes present within any of a parent’s test cases. The results are grouped by the transition route group.

UpdateAgentRequest

The request message for [Agents.UpdateAgent][google.cloud.dialogflow.cx.v3beta1.Agents.UpdateAgent].

UpdateEntityTypeRequest

The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.UpdateEntityType].

UpdateEnvironmentRequest

The request message for [Environments.UpdateEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.UpdateEnvironment].

UpdateExampleRequest

The request message for [Examples.UpdateExample][google.cloud.dialogflow.cx.v3beta1.Examples.UpdateExample].

UpdateExperimentRequest

The request message for [Experiments.UpdateExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.UpdateExperiment].

UpdateFlowRequest

The request message for [Flows.UpdateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.UpdateFlow].

UpdateGenerativeSettingsRequest

Request for [UpdateGenerativeSettings][google.cloud.dialogflow.cx.v3beta1.Agents.UpdateGenerativeSettings] RPC.

UpdateGeneratorRequest

The request message for [Generators.UpdateGenerator][google.cloud.dialogflow.cx.v3beta1.Generators.UpdateGenerator].

UpdateIntentRequest

The request message for [Intents.UpdateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.UpdateIntent].

UpdatePageRequest

The request message for [Pages.UpdatePage][google.cloud.dialogflow.cx.v3beta1.Pages.UpdatePage].

UpdatePlaybookRequest

The request message for [Playbooks.UpdatePlaybook][google.cloud.dialogflow.cx.v3beta1.Playbooks.UpdatePlaybook].

UpdateSecuritySettingsRequest

The request message for [SecuritySettingsService.UpdateSecuritySettings][google.cloud.dialogflow.cx.v3beta1.SecuritySettingsService.UpdateSecuritySettings].

UpdateSessionEntityTypeRequest

The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.UpdateSessionEntityType].

UpdateTestCaseRequest

The request message for [TestCases.UpdateTestCase][google.cloud.dialogflow.cx.v3beta1.TestCases.UpdateTestCase].

UpdateToolRequest

The request message for [Tools.UpdateTool][google.cloud.dialogflow.cx.v3beta1.Tools.UpdateTool].

UpdateTransitionRouteGroupRequest

The request message for [TransitionRouteGroups.UpdateTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.UpdateTransitionRouteGroup].

UpdateVersionRequest

The request message for [Versions.UpdateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.UpdateVersion].

UpdateWebhookRequest

The request message for [Webhooks.UpdateWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.UpdateWebhook].

UserUtterance

UserUtterance represents one message sent by the customer.

ValidateAgentRequest

The request message for [Agents.ValidateAgent][google.cloud.dialogflow.cx.v3beta1.Agents.ValidateAgent].

ValidateFlowRequest

The request message for [Flows.ValidateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ValidateFlow].

ValidationMessage

Agent/flow validation message.

VariantsHistory

The history of variants update.

Version

Represents a version of a flow.

VersionVariants

A list of flow version variants.

VoiceSelectionParams

Description of which voice to use for speech synthesis.

Webhook

Webhooks host the developer’s business logic. During a session, webhooks allow the developer to use the data extracted by Dialogflow’s natural language processing to generate dynamic responses, validate collected data, or trigger actions on the backend.

WebhookRequest

The request message for a webhook call. The request is sent as a JSON object and the field names will be presented in camel cases.

WebhookResponse

The response message for a webhook call.

Enums

AudioEncoding

Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

DataStoreType

Type of a data store. Determines how search is performed in the data store.

ImportStrategy

Import strategies for the conflict resolution of resources (i.e. intents, entities, and webhooks) with identical display names during import operations.

IntentView

Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response.

OutputAudioEncoding

Audio encoding of the output audio format in Text-To-Speech.

OutputState

Output state.

SpeechModelVariant

Variant of the specified [Speech model][google.cloud.dialogflow.cx.v3beta1.InputAudioConfig.model] to use.

SsmlVoiceGender

Gender of the voice as described in SSML voice element.

TestResult

The test result for a test case and an agent environment.