Struct google_api_proto::google::cloud::dialogflow::v2::StreamingAnalyzeContentResponse

pub struct StreamingAnalyzeContentResponse {
    pub recognition_result: Option<StreamingRecognitionResult>,
    pub reply_text: String,
    pub reply_audio: Option<OutputAudio>,
    pub automated_agent_reply: Option<AutomatedAgentReply>,
    pub message: Option<Message>,
    pub human_agent_suggestion_results: Vec<SuggestionResult>,
    pub end_user_suggestion_results: Vec<SuggestionResult>,
    pub dtmf_parameters: Option<DtmfParameters>,
    pub debugging_info: Option<CloudConversationDebuggingInfo>,
}

The top-level message returned from the StreamingAnalyzeContent method.

Multiple response messages can be returned in order:

If the input was set to streaming audio, the first one or more messages contain recognition_result. Each recognition_result represents a more complete transcript of what the user said. The last recognition_result has is_final set to true.

In virtual agent stage: if enable_partial_automated_agent_reply is true, the following N (currently 1 <= N <= 4) messages contain automated_agent_reply and optionally reply_audio returned by the virtual agent. The first (N-1) automated_agent_replys will have automated_agent_reply_type set to PARTIAL. The last automated_agent_reply has automated_agent_reply_type set to FINAL. If enable_partial_automated_agent_reply is not enabled, response stream only contains the final reply.

In human assist stage: the following N (N >= 1) messages contain human_agent_suggestion_results, end_user_suggestion_results or message.

Fields
recognition_result: Option<StreamingRecognitionResult>

The result of speech recognition.

reply_text: String

The output text content. This field is set if an automated agent responded with a text for the user.

reply_audio: Option<OutputAudio>

The audio data bytes encoded as specified in the request. This field is set if:

The reply_audio_config field is specified in the request.

The automated agent, which this output comes from, responded with audio. In such case, the reply_audio.config field contains settings used to synthesize the speech.

In some scenarios, multiple output audio fields may be present in the response structure. In these cases, only the top-most-level audio output has content.

automated_agent_reply: Option<AutomatedAgentReply>

Only set if a Dialogflow automated agent has responded. Note that: [AutomatedAgentReply.detect_intent_response.output_audio][] and [AutomatedAgentReply.detect_intent_response.output_audio_config][] are always empty, use [reply_audio][google.cloud.dialogflow.v2.StreamingAnalyzeContentResponse.reply_audio] instead.

message: Option<Message>

Message analyzed by CCAI.

human_agent_suggestion_results: Vec<SuggestionResult>

The suggestions for most recent human agent. The order is the same as [HumanAgentAssistantConfig.SuggestionConfig.feature_configs][google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionConfig.feature_configs] of [HumanAgentAssistantConfig.human_agent_suggestion_config][google.cloud.dialogflow.v2.HumanAgentAssistantConfig.human_agent_suggestion_config].

end_user_suggestion_results: Vec<SuggestionResult>

The suggestions for end user. The order is the same as [HumanAgentAssistantConfig.SuggestionConfig.feature_configs][google.cloud.dialogflow.v2.HumanAgentAssistantConfig.SuggestionConfig.feature_configs] of [HumanAgentAssistantConfig.end_user_suggestion_config][google.cloud.dialogflow.v2.HumanAgentAssistantConfig.end_user_suggestion_config].

dtmf_parameters: Option<DtmfParameters>

Indicates the parameters of DTMF.

debugging_info: Option<CloudConversationDebuggingInfo>

Debugging info that would get populated when StreamingAnalyzeContentRequest.enable_debugging_info is set to true.