Module google_api_proto::google::cloud::discoveryengine::logging

Modules

import_error_context

Nested message and enum types in ImportErrorContext.

Structs

ErrorContext

A description of the context in which an error occurred.

ErrorLog

An error log which is reported to the Error Reporting system.

HttpRequestContext

HTTP request data that is related to a reported error.

ImportErrorContext

The error payload that is populated on LRO import APIs, including the following:

ServiceContext

Describes a running service that sends errors.

SourceLocation

Indicates a location in the source code of the service for which errors are reported.