Module google_api_proto::google::cloud::discoveryengine::v1alpha

Modules

acl_config_service_client

Generated client implementations.

answer

Nested message and enum types in Answer.

answer_query_request

Nested message and enum types in AnswerQueryRequest.

big_query_source

Nested message and enum types in BigQuerySource.

bigtable_options

Nested message and enum types in BigtableOptions.

check_grounding_response

Nested message and enum types in CheckGroundingResponse.

chunk

Nested message and enum types in Chunk.

chunk_service_client

Generated client implementations.

complete_query_response

Nested message and enum types in CompleteQueryResponse.

completion_service_client

Generated client implementations.

condition

Nested message and enum types in Condition.

control

Nested message and enum types in Control.

control_service_client

Generated client implementations.

conversation

Nested message and enum types in Conversation.

conversation_message

Nested message and enum types in ConversationMessage.

conversational_search_service_client

Generated client implementations.

custom_tuning_model

Nested message and enum types in CustomTuningModel.

data_store

Nested message and enum types in DataStore.

data_store_service_client

Generated client implementations.

document

Nested message and enum types in Document.

document_info

Nested message and enum types in DocumentInfo.

document_processing_config

Nested message and enum types in DocumentProcessingConfig.

document_service_client

Generated client implementations.

engine

Nested message and enum types in Engine.

engine_service_client

Generated client implementations.

estimate_billing_service_client

Generated client implementations.

estimate_data_size_request

Nested message and enum types in EstimateDataSizeRequest.

field_config

Nested message and enum types in FieldConfig.

get_processed_document_request

Nested message and enum types in GetProcessedDocumentRequest.

grounded_generation_service_client

Generated client implementations.

idp_config

Nested message and enum types in IdpConfig.

import_documents_request

Nested message and enum types in ImportDocumentsRequest.

import_error_config

Nested message and enum types in ImportErrorConfig.

import_suggestion_deny_list_entries_request

Nested message and enum types in ImportSuggestionDenyListEntriesRequest.

import_user_events_request

Nested message and enum types in ImportUserEventsRequest.

interval

Nested message and enum types in Interval.

principal

Nested message and enum types in Principal.

processed_document

Nested message and enum types in ProcessedDocument.

project

Nested message and enum types in Project.

project_service_client

Generated client implementations.

purge_documents_request

Nested message and enum types in PurgeDocumentsRequest.

purge_error_config

Nested message and enum types in PurgeErrorConfig.

query

Nested message and enum types in Query.

rank_service_client

Generated client implementations.

recommend_response

Nested message and enum types in RecommendResponse.

recommendation_service_client

Generated client implementations.

recrawl_uris_response

Nested message and enum types in RecrawlUrisResponse.

reply

Nested message and enum types in Reply.

report_consent_change_request

Nested message and enum types in ReportConsentChangeRequest.

schema

Nested message and enum types in Schema.

schema_service_client

Generated client implementations.

search_request

Nested message and enum types in SearchRequest.

search_response

Nested message and enum types in SearchResponse.

search_service_client

Generated client implementations.

search_tuning_service_client

Generated client implementations.

serving_config

Nested message and enum types in ServingConfig.

serving_config_service_client

Generated client implementations.

session

Nested message and enum types in Session.

site_search_engine_service_client

Generated client implementations.

site_verification_info

Nested message and enum types in SiteVerificationInfo.

suggestion_deny_list_entry

Nested message and enum types in SuggestionDenyListEntry.

target_site

Nested message and enum types in TargetSite.

train_custom_model_request

Nested message and enum types in TrainCustomModelRequest.

user_event_service_client

Generated client implementations.

Structs

AclConfig

Access Control Configuration.

Answer

Defines an answer.

AnswerQueryRequest

Request message for [ConversationalSearchService.AnswerQuery][google.cloud.discoveryengine.v1alpha.ConversationalSearchService.AnswerQuery] method.

AnswerQueryResponse

Response message for [ConversationalSearchService.AnswerQuery][google.cloud.discoveryengine.v1alpha.ConversationalSearchService.AnswerQuery] method.

BatchCreateTargetSiteMetadata

Metadata related to the progress of the [SiteSearchEngineService.BatchCreateTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchCreateTargetSites] operation. This will be returned by the google.longrunning.Operation.metadata field.

BatchCreateTargetSitesRequest

Request message for [SiteSearchEngineService.BatchCreateTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchCreateTargetSites] method.

BatchCreateTargetSitesResponse

Response message for [SiteSearchEngineService.BatchCreateTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchCreateTargetSites] method.

BatchVerifyTargetSitesMetadata

Metadata related to the progress of the [SiteSearchEngineService.BatchVerifyTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchVerifyTargetSites] operation. This will be returned by the google.longrunning.Operation.metadata field.

BatchVerifyTargetSitesRequest

Request message for [SiteSearchEngineService.BatchVerifyTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchVerifyTargetSites] method.

BatchVerifyTargetSitesResponse

Response message for [SiteSearchEngineService.BatchVerifyTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.BatchVerifyTargetSites] method.

BigQuerySource

BigQuery source import data from.

BigtableOptions

The Bigtable Options object that contains information to support the import.

BigtableSource

The Cloud Bigtable source for importing data.

CheckGroundingRequest

Request message for [GroundedGenerationService.CheckGrounding][google.cloud.discoveryengine.v1alpha.GroundedGenerationService.CheckGrounding] method.

CheckGroundingResponse

Response message for the [GroundedGenerationService.CheckGrounding][google.cloud.discoveryengine.v1alpha.GroundedGenerationService.CheckGrounding] method.

CheckGroundingSpec

Specification for the grounding check.

Chunk

Chunk captures all raw metadata information of items to be recommended or searched in the chunk mode.

CloudSqlSource

Cloud SQL source import data from.

CollectUserEventRequest

Request message for CollectUserEvent method.

CompleteQueryRequest

Request message for [CompletionService.CompleteQuery][google.cloud.discoveryengine.v1alpha.CompletionService.CompleteQuery] method.

CompleteQueryResponse

Response message for [CompletionService.CompleteQuery][google.cloud.discoveryengine.v1alpha.CompletionService.CompleteQuery] method.

CompletionInfo

Detailed completion information including completion attribution token and clicked completion info.

Condition

Defines circumstances to be checked before allowing a behavior

Control

Defines a conditioned behavior to employ during serving. Must be attached to a [ServingConfig][google.cloud.discoveryengine.v1alpha.ServingConfig] to be considered at serving time. Permitted actions dependent on SolutionType.

Conversation

External conversation proto definition.

ConversationContext

Defines context of the conversation

ConversationMessage

Defines a conversation message.

ConverseConversationRequest

Request message for [ConversationalSearchService.ConverseConversation][google.cloud.discoveryengine.v1alpha.ConversationalSearchService.ConverseConversation] method.

ConverseConversationResponse

Response message for [ConversationalSearchService.ConverseConversation][google.cloud.discoveryengine.v1alpha.ConversationalSearchService.ConverseConversation] method.

CreateControlRequest

Request for CreateControl method.

CreateConversationRequest

Request for CreateConversation method.

CreateDataStoreMetadata

Metadata related to the progress of the [DataStoreService.CreateDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.CreateDataStore] operation. This will be returned by the google.longrunning.Operation.metadata field.

CreateDataStoreRequest

Request for [DataStoreService.CreateDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.CreateDataStore] method.

CreateDocumentRequest

Request message for [DocumentService.CreateDocument][google.cloud.discoveryengine.v1alpha.DocumentService.CreateDocument] method.

CreateEngineMetadata

Metadata related to the progress of the [EngineService.CreateEngine][google.cloud.discoveryengine.v1alpha.EngineService.CreateEngine] operation. This will be returned by the google.longrunning.Operation.metadata field.

CreateEngineRequest

Request for [EngineService.CreateEngine][google.cloud.discoveryengine.v1alpha.EngineService.CreateEngine] method.

CreateSchemaMetadata

Metadata for Create Schema LRO.

CreateSchemaRequest

Request message for [SchemaService.CreateSchema][google.cloud.discoveryengine.v1alpha.SchemaService.CreateSchema] method.

CreateSessionRequest

Request for CreateSession method.

CreateTargetSiteMetadata

Metadata related to the progress of the [SiteSearchEngineService.CreateTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.CreateTargetSite] operation. This will be returned by the google.longrunning.Operation.metadata field.

CreateTargetSiteRequest

Request message for [SiteSearchEngineService.CreateTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.CreateTargetSite] method.

CustomAttribute

A custom attribute that is not explicitly modeled in a resource, e.g. [UserEvent][google.cloud.discoveryengine.v1alpha.UserEvent].

CustomFineTuningSpec

Defines custom fine tuning spec.

CustomTuningModel

Metadata that describes a custom tuned model.

DataStore

DataStore captures global settings and configs at the DataStore level.

DeleteControlRequest

Request for DeleteControl method.

DeleteConversationRequest

Request for DeleteConversation method.

DeleteDataStoreMetadata

Metadata related to the progress of the [DataStoreService.DeleteDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.DeleteDataStore] operation. This will be returned by the google.longrunning.Operation.metadata field.

DeleteDataStoreRequest

Request message for [DataStoreService.DeleteDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.DeleteDataStore] method.

DeleteDocumentRequest

Request message for [DocumentService.DeleteDocument][google.cloud.discoveryengine.v1alpha.DocumentService.DeleteDocument] method.

DeleteEngineMetadata

Metadata related to the progress of the [EngineService.DeleteEngine][google.cloud.discoveryengine.v1alpha.EngineService.DeleteEngine] operation. This will be returned by the google.longrunning.Operation.metadata field.

DeleteEngineRequest

Request message for [EngineService.DeleteEngine][google.cloud.discoveryengine.v1alpha.EngineService.DeleteEngine] method.

DeleteSchemaMetadata

Metadata for DeleteSchema LRO.

DeleteSchemaRequest

Request message for [SchemaService.DeleteSchema][google.cloud.discoveryengine.v1alpha.SchemaService.DeleteSchema] method.

DeleteSessionRequest

Request for DeleteSession method.

DeleteTargetSiteMetadata

Metadata related to the progress of the [SiteSearchEngineService.DeleteTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.DeleteTargetSite] operation. This will be returned by the google.longrunning.Operation.metadata field.

DeleteTargetSiteRequest

Request message for [SiteSearchEngineService.DeleteTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.DeleteTargetSite] method.

DisableAdvancedSiteSearchMetadata

Metadata related to the progress of the [SiteSearchEngineService.DisableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.DisableAdvancedSiteSearch] operation. This will be returned by the google.longrunning.Operation.metadata field.

DisableAdvancedSiteSearchRequest

Request message for [SiteSearchEngineService.DisableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.DisableAdvancedSiteSearch] method.

DisableAdvancedSiteSearchResponse

Response message for [SiteSearchEngineService.DisableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.DisableAdvancedSiteSearch] method.

Document

Document captures all raw metadata information of items to be recommended or searched.

DocumentInfo

Detailed document information associated with a user event.

DocumentProcessingConfig

A singleton resource of [DataStore][google.cloud.discoveryengine.v1alpha.DataStore]. It’s empty when [DataStore][google.cloud.discoveryengine.v1alpha.DataStore] is created, which defaults to digital parser. The first call to [DataStoreService.UpdateDocumentProcessingConfig][google.cloud.discoveryengine.v1alpha.DataStoreService.UpdateDocumentProcessingConfig] method will initialize the config.

DoubleList

Double list.

EmbeddingConfig

Defines embedding config, used for bring your own embeddings feature.

EnableAdvancedSiteSearchMetadata

Metadata related to the progress of the [SiteSearchEngineService.EnableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.EnableAdvancedSiteSearch] operation. This will be returned by the google.longrunning.Operation.metadata field.

EnableAdvancedSiteSearchRequest

Request message for [SiteSearchEngineService.EnableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.EnableAdvancedSiteSearch] method.

EnableAdvancedSiteSearchResponse

Response message for [SiteSearchEngineService.EnableAdvancedSiteSearch][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.EnableAdvancedSiteSearch] method.

Engine

Metadata that describes the training and serving parameters of an [Engine][google.cloud.discoveryengine.v1alpha.Engine].

EstimateDataSizeMetadata

Metadata related to the progress of the EstimateDataSize operation. This is returned by the google.longrunning.Operation.metadata field.

EstimateDataSizeRequest

Request message for [EstimateBillingService.EstimateDataSize][google.cloud.discoveryengine.v1alpha.EstimateBillingService.EstimateDataSize] method

EstimateDataSizeResponse

Response of the EstimateDataSize request. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

FactChunk

Fact Chunk.

FetchDomainVerificationStatusRequest

Request message for [SiteSearchEngineService.FetchDomainVerificationStatus][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.FetchDomainVerificationStatus] method.

FetchDomainVerificationStatusResponse

Response message for [SiteSearchEngineService.FetchDomainVerificationStatus][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.FetchDomainVerificationStatus] method.

FhirStoreSource

Cloud FhirStore source import data from.

FieldConfig

Configurations for fields of a schema. For example, configuring a field is indexable, or searchable.

FirestoreSource

Firestore source import data from.

GcsSource

Cloud Storage location for input content.

GetAclConfigRequest

Request message for GetAclConfigRequest method.

GetAnswerRequest

Request for GetAnswer method.

GetChunkRequest

Request message for [ChunkService.GetChunk][google.cloud.discoveryengine.v1alpha.ChunkService.GetChunk] method.

GetControlRequest

Request for GetControl method.

GetConversationRequest

Request for GetConversation method.

GetDataStoreRequest

Request message for [DataStoreService.GetDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.GetDataStore] method.

GetDocumentProcessingConfigRequest

Request for [DataStoreService.GetDocumentProcessingConfig][google.cloud.discoveryengine.v1alpha.DataStoreService.GetDocumentProcessingConfig] method.

GetDocumentRequest

Request message for [DocumentService.GetDocument][google.cloud.discoveryengine.v1alpha.DocumentService.GetDocument] method.

GetEngineRequest

Request message for [EngineService.GetEngine][google.cloud.discoveryengine.v1alpha.EngineService.GetEngine] method.

GetProcessedDocumentRequest

Request message for [DocumentService.GetDocument][google.cloud.discoveryengine.v1alpha.DocumentService.GetDocument] method.

GetProjectRequest

Request message for [ProjectService.GetProject][google.cloud.discoveryengine.v1alpha.ProjectService.GetProject] method.

GetSchemaRequest

Request message for [SchemaService.GetSchema][google.cloud.discoveryengine.v1alpha.SchemaService.GetSchema] method.

GetServingConfigRequest

Request for GetServingConfig method.

GetSessionRequest

Request for GetSession method.

GetSiteSearchEngineRequest

Request message for [SiteSearchEngineService.GetSiteSearchEngine][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.GetSiteSearchEngine] method.

GetTargetSiteRequest

Request message for [SiteSearchEngineService.GetTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.GetTargetSite] method.

GroundingFact

Grounding Fact.

GuidedSearchSpec

Defines guided search spec.

IdpConfig

Identity Provider Config.

ImportDocumentsMetadata

Metadata related to the progress of the ImportDocuments operation. This is returned by the google.longrunning.Operation.metadata field.

ImportDocumentsRequest

Request message for Import methods.

ImportDocumentsResponse

Response of the [ImportDocumentsRequest][google.cloud.discoveryengine.v1alpha.ImportDocumentsRequest]. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportErrorConfig

Configuration of destination for Import related errors.

ImportSuggestionDenyListEntriesMetadata

Metadata related to the progress of the ImportSuggestionDenyListEntries operation. This is returned by the google.longrunning.Operation.metadata field.

ImportSuggestionDenyListEntriesRequest

Request message for [CompletionService.ImportSuggestionDenyListEntries][google.cloud.discoveryengine.v1alpha.CompletionService.ImportSuggestionDenyListEntries] method.

ImportSuggestionDenyListEntriesResponse

Response message for [CompletionService.ImportSuggestionDenyListEntries][google.cloud.discoveryengine.v1alpha.CompletionService.ImportSuggestionDenyListEntries] method.

ImportUserEventsMetadata

Metadata related to the progress of the Import operation. This is returned by the google.longrunning.Operation.metadata field.

ImportUserEventsRequest

Request message for the ImportUserEvents request.

ImportUserEventsResponse

Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

Interval

A floating point interval.

ListChunksRequest

Request message for [ChunkService.ListChunks][google.cloud.discoveryengine.v1alpha.ChunkService.ListChunks] method.

ListChunksResponse

Response message for [ChunkService.ListChunks][google.cloud.discoveryengine.v1alpha.ChunkService.ListChunks] method.

ListControlsRequest

Request for ListControls method.

ListControlsResponse

Response for ListControls method.

ListConversationsRequest

Request for ListConversations method.

ListConversationsResponse

Response for ListConversations method.

ListCustomModelsRequest

Request message for [SearchTuningService.ListCustomModels][google.cloud.discoveryengine.v1alpha.SearchTuningService.ListCustomModels] method.

ListCustomModelsResponse

Response message for [SearchTuningService.ListCustomModels][google.cloud.discoveryengine.v1alpha.SearchTuningService.ListCustomModels] method.

ListDataStoresRequest

Request message for [DataStoreService.ListDataStores][google.cloud.discoveryengine.v1alpha.DataStoreService.ListDataStores] method.

ListDataStoresResponse

Response message for [DataStoreService.ListDataStores][google.cloud.discoveryengine.v1alpha.DataStoreService.ListDataStores] method.

ListDocumentsRequest

Request message for [DocumentService.ListDocuments][google.cloud.discoveryengine.v1alpha.DocumentService.ListDocuments] method.

ListDocumentsResponse

Response message for [DocumentService.ListDocuments][google.cloud.discoveryengine.v1alpha.DocumentService.ListDocuments] method.

ListEnginesRequest

Request message for [EngineService.ListEngines][google.cloud.discoveryengine.v1alpha.EngineService.ListEngines] method.

ListEnginesResponse

Response message for [EngineService.ListEngines][google.cloud.discoveryengine.v1alpha.EngineService.ListEngines] method.

ListSchemasRequest

Request message for [SchemaService.ListSchemas][google.cloud.discoveryengine.v1alpha.SchemaService.ListSchemas] method.

ListSchemasResponse

Response message for [SchemaService.ListSchemas][google.cloud.discoveryengine.v1alpha.SchemaService.ListSchemas] method.

ListServingConfigsRequest

Request for ListServingConfigs method.

ListServingConfigsResponse

Response for ListServingConfigs method.

ListSessionsRequest

Request for ListSessions method.

ListSessionsResponse

Response for ListSessions method.

ListTargetSitesRequest

Request message for [SiteSearchEngineService.ListTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.ListTargetSites] method.

ListTargetSitesResponse

Response message for [SiteSearchEngineService.ListTargetSites][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.ListTargetSites] method.

MediaInfo

Media-specific user event information.

PageInfo

Detailed page information.

PanelInfo

Detailed panel information associated with a user event.

PauseEngineRequest

Request for pausing training of an engine.

Principal

Principal identifier of a user or a group.

ProcessedDocument

Document captures all raw metadata information of items to be recommended or searched.

Project

Metadata and configurations for a Google Cloud project in the service.

ProvisionProjectMetadata

Metadata associated with a project provision operation.

ProvisionProjectRequest

Request for [ProjectService.ProvisionProject][google.cloud.discoveryengine.v1alpha.ProjectService.ProvisionProject] method.

PurgeDocumentsMetadata

Metadata related to the progress of the PurgeDocuments operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeDocumentsRequest

Request message for [DocumentService.PurgeDocuments][google.cloud.discoveryengine.v1alpha.DocumentService.PurgeDocuments] method.

PurgeDocumentsResponse

Response message for [DocumentService.PurgeDocuments][google.cloud.discoveryengine.v1alpha.DocumentService.PurgeDocuments] method. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

PurgeErrorConfig

Configuration of destination for Purge related errors.

PurgeSuggestionDenyListEntriesMetadata

Metadata related to the progress of the PurgeSuggestionDenyListEntries operation. This is returned by the google.longrunning.Operation.metadata field.

PurgeSuggestionDenyListEntriesRequest

Request message for [CompletionService.PurgeSuggestionDenyListEntries][google.cloud.discoveryengine.v1alpha.CompletionService.PurgeSuggestionDenyListEntries] method.

PurgeSuggestionDenyListEntriesResponse

Response message for [CompletionService.PurgeSuggestionDenyListEntries][google.cloud.discoveryengine.v1alpha.CompletionService.PurgeSuggestionDenyListEntries] method.

PurgeUserEventsMetadata

Metadata related to the progress of the PurgeUserEvents operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeUserEventsRequest

Request message for PurgeUserEvents method.

PurgeUserEventsResponse

Response of the PurgeUserEventsRequest. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

Query

Defines a user inputed query.

RankRequest

Request message for [RankService.Rank][google.cloud.discoveryengine.v1alpha.RankService.Rank] method.

RankResponse

Response message for [RankService.Rank][google.cloud.discoveryengine.v1alpha.RankService.Rank] method.

RankingRecord

Record message for [RankService.Rank][google.cloud.discoveryengine.v1alpha.RankService.Rank] method.

RecommendRequest

Request message for Recommend method.

RecommendResponse

Response message for Recommend method.

RecrawlUrisMetadata

Metadata related to the progress of the [SiteSearchEngineService.RecrawlUris][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.RecrawlUris] operation. This will be returned by the google.longrunning.Operation.metadata field.

RecrawlUrisRequest

Request message for [SiteSearchEngineService.RecrawlUris][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.RecrawlUris] method.

RecrawlUrisResponse

Response message for [SiteSearchEngineService.RecrawlUris][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.RecrawlUris] method.

Reply

Defines a reply message to user.

ReportConsentChangeRequest

Request for ReportConsentChange method.

ResumeEngineRequest

Request for resuming training of an engine.

Schema

Defines the structure and layout of a type of document data.

SearchInfo

Detailed search information.

SearchRequest

Request message for [SearchService.Search][google.cloud.discoveryengine.v1alpha.SearchService.Search] method.

SearchResponse

Response message for [SearchService.Search][google.cloud.discoveryengine.v1alpha.SearchService.Search] method.

ServingConfig

Configures metadata that is used to generate serving time results (e.g. search results or recommendation predictions). The ServingConfig is passed in the search and predict request and generates results.

Session

External session proto definition.

SiteSearchEngine

SiteSearchEngine captures DataStore level site search persisting configurations. It is a singleton value per data store.

SiteVerificationInfo

Verification information for target sites in advanced site search.

SpannerSource

The Spanner source for importing data

SuggestionDenyListEntry

Suggestion deny list entry identifying the phrase to block from suggestions and the applied operation for the phrase.

TargetSite

A target site for the SiteSearchEngine.

TextInput

Defines text input.

TrainCustomModelMetadata

Metadata related to the progress of the TrainCustomModel operation. This is returned by the google.longrunning.Operation.metadata field.

TrainCustomModelRequest

Request message for [SearchTuningService.TrainCustomModel][google.cloud.discoveryengine.v1alpha.SearchTuningService.TrainCustomModel] method.

TrainCustomModelResponse

Response of the [TrainCustomModelRequest][google.cloud.discoveryengine.v1alpha.TrainCustomModelRequest]. This message is returned by the google.longrunning.Operations.response field.

TransactionInfo

A transaction represents the entire purchase transaction.

TuneEngineMetadata

Metadata associated with a tune operation.

TuneEngineRequest

Request to manually start a tuning process now (instead of waiting for the periodically scheduled tuning to happen).

TuneEngineResponse

Response associated with a tune operation.

UpdateAclConfigRequest

Request message for UpdateAclConfig method.

UpdateControlRequest

Request for UpdateControl method.

UpdateConversationRequest

Request for UpdateConversation method.

UpdateDataStoreRequest

Request message for [DataStoreService.UpdateDataStore][google.cloud.discoveryengine.v1alpha.DataStoreService.UpdateDataStore] method.

UpdateDocumentProcessingConfigRequest

Request for [DataStoreService.UpdateDocumentProcessingConfig][google.cloud.discoveryengine.v1alpha.DataStoreService.UpdateDocumentProcessingConfig] method.

UpdateDocumentRequest

Request message for [DocumentService.UpdateDocument][google.cloud.discoveryengine.v1alpha.DocumentService.UpdateDocument] method.

UpdateEngineRequest

Request message for [EngineService.UpdateEngine][google.cloud.discoveryengine.v1alpha.EngineService.UpdateEngine] method.

UpdateSchemaMetadata

Metadata for UpdateSchema LRO.

UpdateSchemaRequest

Request message for [SchemaService.UpdateSchema][google.cloud.discoveryengine.v1alpha.SchemaService.UpdateSchema] method.

UpdateServingConfigRequest

Request for UpdateServingConfig method.

UpdateSessionRequest

Request for UpdateSession method.

UpdateTargetSiteMetadata

Metadata related to the progress of the [SiteSearchEngineService.UpdateTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.UpdateTargetSite] operation. This will be returned by the google.longrunning.Operation.metadata field.

UpdateTargetSiteRequest

Request message for [SiteSearchEngineService.UpdateTargetSite][google.cloud.discoveryengine.v1alpha.SiteSearchEngineService.UpdateTargetSite] method.

UserEvent

UserEvent captures all metadata information Discovery Engine API needs to know about how end users interact with your website.

UserInfo

Information of an end user.

WriteUserEventRequest

Request message for WriteUserEvent method.

Enums

IndustryVertical

The industry vertical associated with the [DataStore][google.cloud.discoveryengine.v1alpha.DataStore].

SearchAddOn

Add-on that provides additional functionality for search.

SearchTier

Tiers of search features. Different tiers might have different pricing. To learn more, check the pricing documentation.

SearchUseCase

Defines a further subdivision of SolutionType. Specifically applies to [SOLUTION_TYPE_SEARCH][google.cloud.discoveryengine.v1alpha.SolutionType.SOLUTION_TYPE_SEARCH].

SolutionType

The type of solution.