Struct google_api_proto::google::cloud::discoveryengine::v1alpha::RecommendRequest

pub struct RecommendRequest {
    pub serving_config: String,
    pub user_event: Option<UserEvent>,
    pub page_size: i32,
    pub filter: String,
    pub validate_only: bool,
    pub params: BTreeMap<String, Value>,
    pub user_labels: BTreeMap<String, String>,
}

Request message for Recommend method.

Fields
serving_config: String

Required. Full resource name of a [ServingConfig][google.cloud.discoveryengine.v1alpha.ServingConfig]: projects/*/locations/global/collections/*/engines/*/servingConfigs/*, or projects/*/locations/global/collections/*/dataStores/*/servingConfigs/*

One default serving config is created along with your recommendation engine creation. The engine ID is used as the ID of the default serving config. For example, for Engine projects/*/locations/global/collections/*/engines/my-engine, you can use projects/*/locations/global/collections/*/engines/my-engine/servingConfigs/my-engine for your [RecommendationService.Recommend][google.cloud.discoveryengine.v1alpha.RecommendationService.Recommend] requests.

user_event: Option<UserEvent>

Required. Context about the user, what they are looking at and what action they took to trigger the Recommend request. Note that this user event detail won’t be ingested to userEvent logs. Thus, a separate userEvent write request is required for event logging.

Don’t set [UserEvent.user_pseudo_id][google.cloud.discoveryengine.v1alpha.UserEvent.user_pseudo_id] or [UserEvent.user_info.user_id][google.cloud.discoveryengine.v1alpha.UserInfo.user_id] to the same fixed ID for different users. If you are trying to receive non-personalized recommendations (not recommended; this can negatively impact model performance), instead set [UserEvent.user_pseudo_id][google.cloud.discoveryengine.v1alpha.UserEvent.user_pseudo_id] to a random unique ID and leave [UserEvent.user_info.user_id][google.cloud.discoveryengine.v1alpha.UserInfo.user_id] unset.

page_size: i32

Maximum number of results to return. Set this property to the number of recommendation results needed. If zero, the service chooses a reasonable default. The maximum allowed value is 100. Values above 100 are set to 100.

filter: String

Filter for restricting recommendation results with a length limit of 5,000 characters. Currently, only filter expressions on the filter_tags attribute is supported.

Examples:

(filter_tags: ANY("Red", "Blue") OR filter_tags: ANY("Hot", "Cold"))

(filter_tags: ANY("Red", "Blue")) AND NOT (filter_tags: ANY("Green"))

If attributeFilteringSyntax is set to true under the params field, then attribute-based expressions are expected instead of the above described tag-based syntax. Examples:

(launguage: ANY(“en”, “es”)) AND NOT (categories: ANY(“Movie”))

(available: true) AND (launguage: ANY(“en”, “es”)) OR (categories: ANY(“Movie”))

If your filter blocks all results, the API returns generic (unfiltered) popular Documents. If you only want results strictly matching the filters, set strictFiltering to true in [RecommendRequest.params][google.cloud.discoveryengine.v1alpha.RecommendRequest.params] to receive empty results instead.

Note that the API never returns [Document][google.cloud.discoveryengine.v1alpha.Document]s with storageStatus as EXPIRED or DELETED regardless of filter choices.

validate_only: bool

Use validate only mode for this recommendation query. If set to true, a fake model is used that returns arbitrary Document IDs. Note that the validate only mode should only be used for testing the API, or if the model is not ready.

params: BTreeMap<String, Value>

Additional domain specific parameters for the recommendations.

Allowed values:

returnDocument: Boolean. If set to true, the associated Document object is returned in [RecommendResponse.RecommendationResult.document][google.cloud.discoveryengine.v1alpha.RecommendResponse.RecommendationResult.document].

returnScore: Boolean. If set to true, the recommendation score corresponding to each returned Document is set in [RecommendResponse.RecommendationResult.metadata][google.cloud.discoveryengine.v1alpha.RecommendResponse.RecommendationResult.metadata]. The given score indicates the probability of a Document conversion given the user’s context and history.

strictFiltering: Boolean. True by default. If set to false, the service returns generic (unfiltered) popular Documents instead of empty if your filter blocks all recommendation results.

diversityLevel: String. Default empty. If set to be non-empty, then it needs to be one of:

no-diversity

low-diversity

medium-diversity

high-diversity

auto-diversity This gives request-level control and adjusts recommendation results based on Document category.

attributeFilteringSyntax: Boolean. False by default. If set to true, the filter field is interpreted according to the new, attribute-based syntax.

user_labels: BTreeMap<String, String>

The user labels applied to a resource must meet the following requirements:

Each resource can have multiple labels, up to a maximum of 64.

Each label must be a key-value pair.

Keys have a minimum length of 1 character and a maximum length of 63 characters and cannot be empty. Values can be empty and have a maximum length of 63 characters.

Keys and values can contain only lowercase letters, numeric characters, underscores, and dashes. All characters must use UTF-8 encoding, and international characters are allowed.

The key portion of a label must be unique. However, you can use the same key with multiple resources.

Keys must start with a lowercase letter or international character.

See Requirements for labels for more details.