Module google_api_proto::google::cloud::discoveryengine::v1beta::search_request

Nested message and enum types in SearchRequest.

Modules

boost_spec

Nested message and enum types in BoostSpec.

content_search_spec

Nested message and enum types in ContentSearchSpec.

embedding_spec

Nested message and enum types in EmbeddingSpec.

facet_spec

Nested message and enum types in FacetSpec.

image_query

Nested message and enum types in ImageQuery.

natural_language_query_understanding_spec

Nested message and enum types in NaturalLanguageQueryUnderstandingSpec.

query_expansion_spec

Nested message and enum types in QueryExpansionSpec.

search_as_you_type_spec

Nested message and enum types in SearchAsYouTypeSpec.

spell_correction_spec

Nested message and enum types in SpellCorrectionSpec.

Structs

BoostSpec

Boost specification to boost certain documents.

ContentSearchSpec

A specification for configuring the behavior of content search.

DataStoreSpec

A struct to define data stores to filter on in a search call and configurations for those data stores. Otherwise, an INVALID_ARGUMENT error is returned.

EmbeddingSpec

The specification that uses customized query embedding vector to do semantic document retrieval.

FacetSpec

A facet specification to perform faceted search.

ImageQuery

Specifies the image query input.

NaturalLanguageQueryUnderstandingSpec

Specification to enable natural language understanding capabilities for search requests.

QueryExpansionSpec

Specification to determine under which conditions query expansion should occur.

SearchAsYouTypeSpec

Specification for search as you type in search requests.

SessionSpec

Session specification.

SpellCorrectionSpec

The specification for query spell correction.