Struct google_api_proto::google::cloud::discoveryengine::v1beta::search_request::content_search_spec::SummarySpec

pub struct SummarySpec {
    pub summary_result_count: i32,
    pub include_citations: bool,
    pub ignore_adversarial_query: bool,
    pub ignore_non_summary_seeking_query: bool,
    pub model_prompt_spec: Option<ModelPromptSpec>,
    pub language_code: String,
    pub model_spec: Option<ModelSpec>,
    pub use_semantic_chunks: bool,
}

A specification for configuring a summary returned in a search response.

Fields
summary_result_count: i32

The number of top results to generate the summary from. If the number of results returned is less than summaryResultCount, the summary is generated from all of the results.

At most 10 results for documents mode, or 50 for chunks mode, can be used to generate a summary. The chunks mode is used when [SearchRequest.ContentSearchSpec.search_result_mode][] is set to [CHUNKS][SearchRequest.ContentSearchSpec.SearchResultMode.CHUNKS].

include_citations: bool

Specifies whether to include citations in the summary. The default value is false.

When this field is set to true, summaries include in-line citation numbers.

Example summary including citations:

BigQuery is Google Cloud’s fully managed and completely serverless enterprise data warehouse [1]. BigQuery supports all data types, works across clouds, and has built-in machine learning and business intelligence, all within a unified platform [2, 3].

The citation numbers refer to the returned search results and are 1-indexed. For example, [1] means that the sentence is attributed to the first search result. [2, 3] means that the sentence is attributed to both the second and third search results.

ignore_adversarial_query: bool

Specifies whether to filter out adversarial queries. The default value is false.

Google employs search-query classification to detect adversarial queries. No summary is returned if the search query is classified as an adversarial query. For example, a user might ask a question regarding negative comments about the company or submit a query designed to generate unsafe, policy-violating output. If this field is set to true, we skip generating summaries for adversarial queries and return fallback messages instead.

ignore_non_summary_seeking_query: bool

Specifies whether to filter out queries that are not summary-seeking. The default value is false.

Google employs search-query classification to detect summary-seeking queries. No summary is returned if the search query is classified as a non-summary seeking query. For example, why is the sky blue and Who is the best soccer player in the world? are summary-seeking queries, but SFO airport and world cup 2026 are not. They are most likely navigational queries. If this field is set to true, we skip generating summaries for non-summary seeking queries and return fallback messages instead.

model_prompt_spec: Option<ModelPromptSpec>

If specified, the spec will be used to modify the prompt provided to the LLM.

language_code: String

Language code for Summary. Use language tags defined by BCP47. Note: This is an experimental feature.

model_spec: Option<ModelSpec>

If specified, the spec will be used to modify the model specification provided to the LLM.

use_semantic_chunks: bool

If true, answer will be generated from most relevant chunks from top search results. This feature will improve summary quality. Note that with this feature enabled, not all top search results will be referenced and included in the reference list, so the citation source index only points to the search results listed in the reference list.