Struct google_api_proto::google::cloud::discoveryengine::v1beta::SearchRequest

pub struct SearchRequest {

    pub serving_config: String,
    pub branch: String,
    pub query: String,
    pub image_query: Option<ImageQuery>,
    pub page_size: i32,
    pub page_token: String,
    pub offset: i32,
    pub data_store_specs: Vec<DataStoreSpec>,
    pub filter: String,
    pub canonical_filter: String,
    pub order_by: String,
    pub user_info: Option<UserInfo>,
    pub language_code: String,
    pub region_code: String,
    pub facet_specs: Vec<FacetSpec>,
    pub boost_spec: Option<BoostSpec>,
    pub params: BTreeMap<String, Value>,
    pub query_expansion_spec: Option<QueryExpansionSpec>,
    pub spell_correction_spec: Option<SpellCorrectionSpec>,
    pub user_pseudo_id: String,
    pub content_search_spec: Option<ContentSearchSpec>,
    pub embedding_spec: Option<EmbeddingSpec>,
    pub ranking_expression: String,
    pub safe_search: bool,
    pub user_labels: BTreeMap<String, String>,
    pub natural_language_query_understanding_spec: Option<NaturalLanguageQueryUnderstandingSpec>,
    pub search_as_you_type_spec: Option<SearchAsYouTypeSpec>,
    pub session: String,
    pub session_spec: Option<SessionSpec>,

}

Request message for [SearchService.Search][google.cloud.discoveryengine.v1beta.SearchService.Search] method.

Fields
serving_config: String

Required. The resource name of the Search serving config, such as projects/*/locations/global/collections/default_collection/engines/*/servingConfigs/default_serving_config, or projects/*/locations/global/collections/default_collection/dataStores/default_data_store/servingConfigs/default_serving_config. This field is used to identify the serving configuration name, set of models used to make the search.

branch: String

The branch resource name, such as projects/*/locations/global/collections/default_collection/dataStores/default_data_store/branches/0.

Use default_branch as the branch ID or leave this field empty, to search documents under the default branch.

query: String

Raw search query.

image_query: Option<ImageQuery>

Raw image query.

page_size: i32

Maximum number of [Document][google.cloud.discoveryengine.v1beta.Document]s to return. The maximum allowed value depends on the data type. Values above the maximum value are coerced to the maximum value.

Websites with basic indexing: Default 10, Maximum 25.

Websites with advanced indexing: Default 25, Maximum 50.

Other: Default 50, Maximum 100.

If this field is negative, an INVALID_ARGUMENT is returned.

page_token: String

A page token received from a previous [SearchService.Search][google.cloud.discoveryengine.v1beta.SearchService.Search] call. Provide this to retrieve the subsequent page.

When paginating, all other parameters provided to [SearchService.Search][google.cloud.discoveryengine.v1beta.SearchService.Search] must match the call that provided the page token. Otherwise, an INVALID_ARGUMENT error is returned.

offset: i32

A 0-indexed integer that specifies the current offset (that is, starting result location, amongst the [Document][google.cloud.discoveryengine.v1beta.Document]s deemed by the API as relevant) in search results. This field is only considered if [page_token][google.cloud.discoveryengine.v1beta.SearchRequest.page_token] is unset.

If this field is negative, an INVALID_ARGUMENT is returned.

data_store_specs: Vec<DataStoreSpec>

Specs defining dataStores to filter on in a search call and configurations for those dataStores. This is only considered for engines with multiple dataStores use case. For single dataStore within an engine, they should use the specs at the top level.

filter: String

The filter syntax consists of an expression language for constructing a predicate from one or more fields of the documents being filtered. Filter expression is case-sensitive.

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

Filtering in Vertex AI Search is done by mapping the LHS filter key to a key property defined in the Vertex AI Search backend – this mapping is defined by the customer in their schema. For example a media customer might have a field ‘name’ in their schema. In this case the filter would look like this: filter –> name:‘ANY(“king kong”)’

For more information about filtering including syntax and filter operators, see Filter

canonical_filter: String

The default filter that is applied when a user performs a search without checking any filters on the search page.

The filter applied to every search request when quality improvement such as query expansion is needed. In the case a query does not have a sufficient amount of results this filter will be used to determine whether or not to enable the query expansion flow. The original filter will still be used for the query expanded search. This field is strongly recommended to achieve high search quality.

For more information about filter syntax, see [SearchRequest.filter][google.cloud.discoveryengine.v1beta.SearchRequest.filter].

order_by: String

The order in which documents are returned. Documents can be ordered by a field in an [Document][google.cloud.discoveryengine.v1beta.Document] object. Leave it unset if ordered by relevance. order_by expression is case-sensitive.

For more information on ordering for retail search, see Ordering

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

user_info: Option<UserInfo>

Information about the end user. Highly recommended for analytics. [UserInfo.user_agent][google.cloud.discoveryengine.v1beta.UserInfo.user_agent] is used to deduce device_type for analytics.

language_code: String

The BCP-47 language code, such as “en-US” or “sr-Latn”. For more information, see Standard fields. This field helps to better interpret the query. If a value isn’t specified, the query language code is automatically detected, which may not be accurate.

region_code: String

The Unicode country/region code (CLDR) of a location, such as “US” and “419”. For more information, see Standard fields. If set, then results will be boosted based on the region_code provided.

facet_specs: Vec<FacetSpec>

Facet specifications for faceted search. If empty, no facets are returned.

A maximum of 100 values are allowed. Otherwise, an INVALID_ARGUMENT error is returned.

boost_spec: Option<BoostSpec>

Boost specification to boost certain documents. For more information on boosting, see Boosting

params: BTreeMap<String, Value>

Additional search parameters.

For public website search only, supported values are:

user_country_code: string. Default empty. If set to non-empty, results are restricted or boosted based on the location provided. For example, user_country_code: "au"

For available codes see Country Codes

search_type: double. Default empty. Enables non-webpage searching depending on the value. The only valid non-default value is 1, which enables image searching. For example, search_type: 1

query_expansion_spec: Option<QueryExpansionSpec>

The query expansion specification that specifies the conditions under which query expansion occurs.

spell_correction_spec: Option<SpellCorrectionSpec>

The spell correction specification that specifies the mode under which spell correction takes effect.

user_pseudo_id: String

A unique identifier for tracking visitors. For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor logs in or out of the website.

This field should NOT have a fixed value such as unknown_visitor.

This should be the same identifier as [UserEvent.user_pseudo_id][google.cloud.discoveryengine.v1beta.UserEvent.user_pseudo_id] and [CompleteQueryRequest.user_pseudo_id][google.cloud.discoveryengine.v1beta.CompleteQueryRequest.user_pseudo_id]

The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.

content_search_spec: Option<ContentSearchSpec>

A specification for configuring the behavior of content search.

embedding_spec: Option<EmbeddingSpec>

Uses the provided embedding to do additional semantic document retrieval. The retrieval is based on the dot product of [SearchRequest.EmbeddingSpec.EmbeddingVector.vector][google.cloud.discoveryengine.v1beta.SearchRequest.EmbeddingSpec.EmbeddingVector.vector] and the document embedding that is provided in [SearchRequest.EmbeddingSpec.EmbeddingVector.field_path][google.cloud.discoveryengine.v1beta.SearchRequest.EmbeddingSpec.EmbeddingVector.field_path].

If [SearchRequest.EmbeddingSpec.EmbeddingVector.field_path][google.cloud.discoveryengine.v1beta.SearchRequest.EmbeddingSpec.EmbeddingVector.field_path] is not provided, it will use [ServingConfig.EmbeddingConfig.field_path][google.cloud.discoveryengine.v1beta.ServingConfig.embedding_config].

ranking_expression: String

The ranking expression controls the customized ranking on retrieval documents. This overrides [ServingConfig.ranking_expression][google.cloud.discoveryengine.v1beta.ServingConfig.ranking_expression]. The ranking expression is a single function or multiple functions that are joined by “+”.

ranking_expression = function, { “ + “, function };

Supported functions:

double * relevance_score

double * dotProduct(embedding_field_path)

Function variables:

relevance_score: pre-defined keywords, used for measure relevance between query and document.

embedding_field_path: the document embedding field used with query embedding vector.

dotProduct: embedding function between embedding_field_path and query embedding vector.

Example ranking expression:

If document has an embedding field doc_embedding, the ranking expression
could be `0.5 * relevance_score + 0.3 * dotProduct(doc_embedding)`.

safe_search: bool

Whether to turn on safe search. This is only supported for website search.

user_labels: BTreeMap<String, String>

The user labels applied to a resource must meet the following requirements:

Each resource can have multiple labels, up to a maximum of 64.

Each label must be a key-value pair.

Keys have a minimum length of 1 character and a maximum length of 63 characters and cannot be empty. Values can be empty and have a maximum length of 63 characters.

Keys and values can contain only lowercase letters, numeric characters, underscores, and dashes. All characters must use UTF-8 encoding, and international characters are allowed.

The key portion of a label must be unique. However, you can use the same key with multiple resources.

Keys must start with a lowercase letter or international character.

See Google Cloud Document for more details.

natural_language_query_understanding_spec: Option<NaturalLanguageQueryUnderstandingSpec>

If naturalLanguageQueryUnderstandingSpec is not specified, no additional natural language query understanding will be done.

search_as_you_type_spec: Option<SearchAsYouTypeSpec>

Search as you type configuration. Only supported for the [IndustryVertical.MEDIA][google.cloud.discoveryengine.v1beta.IndustryVertical.MEDIA] vertical.

session: String

The session resource name. Optional.

Session allows users to do multi-turn /search API calls or coordination between /search API calls and /answer API calls.

Example #1 (multi-turn /search API calls):

Call /search API with the auto-session mode (see below).

Call /search API with the session ID generated in the first call. Here, the previous search query gets considered in query standing. I.e., if the first query is “How did Alphabet do in 2022?” and the current query is “How about 2023?”, the current query will be interpreted as “How did Alphabet do in 2023?”.

Example #2 (coordination between /search API calls and /answer API calls):

Call /search API with the auto-session mode (see below).

Call /answer API with the session ID generated in the first call. Here, the answer generation happens in the context of the search results from the first search call.

Auto-session mode: when projects/.../sessions/- is used, a new session gets automatically created. Otherwise, users can use the create-session API to create a session manually.

Multi-turn Search feature is currently at private GA stage. Please use v1alpha or v1beta version instead before we launch this feature to public GA. Or ask for allowlisting through Google Support team.

session_spec: Option<SessionSpec>

Session specification.

Can be used only when session is set.