Module google_api_proto::google::cloud::documentai::v1

Modules

batch_documents_input_config

Nested message and enum types in BatchDocumentsInputConfig.

batch_process_metadata

Nested message and enum types in BatchProcessMetadata.

common_operation_metadata

Nested message and enum types in CommonOperationMetadata.

document

Nested message and enum types in Document.

document_output_config

Nested message and enum types in DocumentOutputConfig.

document_processor_service_client

Generated client implementations.

document_schema

Nested message and enum types in DocumentSchema.

evaluation

Nested message and enum types in Evaluation.

human_review_status

Nested message and enum types in HumanReviewStatus.

ocr_config

Nested message and enum types in OcrConfig.

process_options

Nested message and enum types in ProcessOptions.

process_request

Nested message and enum types in ProcessRequest.

processor

Nested message and enum types in Processor.

processor_type

Nested message and enum types in ProcessorType.

processor_version

Nested message and enum types in ProcessorVersion.

review_document_request

Nested message and enum types in ReviewDocumentRequest.

review_document_response

Nested message and enum types in ReviewDocumentResponse.

train_processor_version_metadata

Nested message and enum types in TrainProcessorVersionMetadata.

train_processor_version_request

Nested message and enum types in TrainProcessorVersionRequest.

Structs

Barcode

Encodes the detailed information of a barcode.

BatchDocumentsInputConfig

The common config to specify a set of documents used as input.

BatchProcessMetadata

The long-running operation metadata for [BatchProcessDocuments][google.cloud.documentai.v1.DocumentProcessorService.BatchProcessDocuments].

BatchProcessRequest

Request message for [BatchProcessDocuments][google.cloud.documentai.v1.DocumentProcessorService.BatchProcessDocuments].

BatchProcessResponse

Response message for [BatchProcessDocuments][google.cloud.documentai.v1.DocumentProcessorService.BatchProcessDocuments].

BoundingPoly

A bounding polygon for the detected image annotation.

CommonOperationMetadata

The common metadata for long running operations.

CreateProcessorRequest

Request message for the [CreateProcessor][google.cloud.documentai.v1.DocumentProcessorService.CreateProcessor] method. Notice this request is sent to a regionalized backend service. If the [ProcessorType][google.cloud.documentai.v1.ProcessorType] isn’t available in that region, the creation fails.

DeleteProcessorMetadata

The long-running operation metadata for the [DeleteProcessor][google.cloud.documentai.v1.DocumentProcessorService.DeleteProcessor] method.

DeleteProcessorRequest

Request message for the [DeleteProcessor][google.cloud.documentai.v1.DocumentProcessorService.DeleteProcessor] method.

DeleteProcessorVersionMetadata

The long-running operation metadata for the [DeleteProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.DeleteProcessorVersion] method.

DeleteProcessorVersionRequest

Request message for the [DeleteProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.DeleteProcessorVersion] method.

DeployProcessorVersionMetadata

The long-running operation metadata for the [DeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.DeployProcessorVersion] method.

DeployProcessorVersionRequest

Request message for the [DeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.DeployProcessorVersion] method.

DeployProcessorVersionResponse

Response message for the [DeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.DeployProcessorVersion] method.

DisableProcessorMetadata

The long-running operation metadata for the [DisableProcessor][google.cloud.documentai.v1.DocumentProcessorService.DisableProcessor] method.

DisableProcessorRequest

Request message for the [DisableProcessor][google.cloud.documentai.v1.DocumentProcessorService.DisableProcessor] method.

DisableProcessorResponse

Response message for the [DisableProcessor][google.cloud.documentai.v1.DocumentProcessorService.DisableProcessor] method. Intentionally empty proto for adding fields in future.

Document

Document represents the canonical document resource in Document AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document AI to iterate and optimize for quality.

DocumentOutputConfig

Config that controls the output of documents. All documents will be written as a JSON file.

DocumentSchema

The schema defines the output of the processed document by a processor.

EnableProcessorMetadata

The long-running operation metadata for the [EnableProcessor][google.cloud.documentai.v1.DocumentProcessorService.EnableProcessor] method.

EnableProcessorRequest

Request message for the [EnableProcessor][google.cloud.documentai.v1.DocumentProcessorService.EnableProcessor] method.

EnableProcessorResponse

Response message for the [EnableProcessor][google.cloud.documentai.v1.DocumentProcessorService.EnableProcessor] method. Intentionally empty proto for adding fields in future.

EvaluateProcessorVersionMetadata

Metadata of the [EvaluateProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.EvaluateProcessorVersion] method.

EvaluateProcessorVersionRequest

Evaluates the given [ProcessorVersion][google.cloud.documentai.v1.ProcessorVersion] against the supplied documents.

EvaluateProcessorVersionResponse

Response of the [EvaluateProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.EvaluateProcessorVersion] method.

Evaluation

An evaluation of a ProcessorVersion’s performance.

EvaluationReference

Gives a short summary of an evaluation, and links to the evaluation itself.

FetchProcessorTypesRequest

Request message for the [FetchProcessorTypes][google.cloud.documentai.v1.DocumentProcessorService.FetchProcessorTypes] method. Some processor types may require the project be added to an allowlist.

FetchProcessorTypesResponse

Response message for the [FetchProcessorTypes][google.cloud.documentai.v1.DocumentProcessorService.FetchProcessorTypes] method.

GcsDocument

Specifies a document stored on Cloud Storage.

GcsDocuments

Specifies a set of documents on Cloud Storage.

GcsPrefix

Specifies all documents on Cloud Storage with a common prefix.

GetEvaluationRequest

Retrieves a specific Evaluation.

GetProcessorRequest

Request message for the [GetProcessor][google.cloud.documentai.v1.DocumentProcessorService.GetProcessor] method.

GetProcessorTypeRequest

Request message for the [GetProcessorType][google.cloud.documentai.v1.DocumentProcessorService.GetProcessorType] method.

GetProcessorVersionRequest

Request message for the [GetProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.GetProcessorVersion] method.

HumanReviewStatus

The status of human review on a processed document.

ListEvaluationsRequest

Retrieves a list of evaluations for a given [ProcessorVersion][google.cloud.documentai.v1.ProcessorVersion].

ListEvaluationsResponse

The response from ListEvaluations.

ListProcessorTypesRequest

Request message for the [ListProcessorTypes][google.cloud.documentai.v1.DocumentProcessorService.ListProcessorTypes] method. Some processor types may require the project be added to an allowlist.

ListProcessorTypesResponse

Response message for the [ListProcessorTypes][google.cloud.documentai.v1.DocumentProcessorService.ListProcessorTypes] method.

ListProcessorVersionsRequest

Request message for list all processor versions belongs to a processor.

ListProcessorVersionsResponse

Response message for the [ListProcessorVersions][google.cloud.documentai.v1.DocumentProcessorService.ListProcessorVersions] method.

ListProcessorsRequest

Request message for list all processors belongs to a project.

ListProcessorsResponse

Response message for the [ListProcessors][google.cloud.documentai.v1.DocumentProcessorService.ListProcessors] method.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

OcrConfig

Config for Document OCR.

ProcessOptions

Options for Process API

ProcessRequest

Request message for the [ProcessDocument][google.cloud.documentai.v1.DocumentProcessorService.ProcessDocument] method.

ProcessResponse

Response message for the [ProcessDocument][google.cloud.documentai.v1.DocumentProcessorService.ProcessDocument] method.

Processor

The first-class citizen for Document AI. Each processor defines how to extract structural information from a document.

ProcessorType

A processor type is responsible for performing a certain document understanding task on a certain type of document.

ProcessorVersion

A processor version is an implementation of a processor. Each processor can have multiple versions, pretrained by Google internally or uptrained by the customer. A processor can only have one default version at a time. Its document-processing behavior is defined by that version.

ProcessorVersionAlias

Contains the alias and the aliased resource name of processor version.

RawDocument

Payload message of raw document content (bytes).

ReviewDocumentOperationMetadata

The long-running operation metadata for the [ReviewDocument][google.cloud.documentai.v1.DocumentProcessorService.ReviewDocument] method.

ReviewDocumentRequest

Request message for the [ReviewDocument][google.cloud.documentai.v1.DocumentProcessorService.ReviewDocument] method.

ReviewDocumentResponse

Response message for the [ReviewDocument][google.cloud.documentai.v1.DocumentProcessorService.ReviewDocument] method.

SetDefaultProcessorVersionMetadata

The long-running operation metadata for the [SetDefaultProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.SetDefaultProcessorVersion] method.

SetDefaultProcessorVersionRequest

Request message for the [SetDefaultProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.SetDefaultProcessorVersion] method.

SetDefaultProcessorVersionResponse

Response message for the [SetDefaultProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.SetDefaultProcessorVersion] method.

TrainProcessorVersionMetadata

The metadata that represents a processor version being created.

TrainProcessorVersionRequest

Request message for the [TrainProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.TrainProcessorVersion] method.

TrainProcessorVersionResponse

The response for [TrainProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.TrainProcessorVersion].

UndeployProcessorVersionMetadata

The long-running operation metadata for the [UndeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.UndeployProcessorVersion] method.

UndeployProcessorVersionRequest

Request message for the [UndeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.UndeployProcessorVersion] method.

UndeployProcessorVersionResponse

Response message for the [UndeployProcessorVersion][google.cloud.documentai.v1.DocumentProcessorService.UndeployProcessorVersion] method.

Vertex

A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.