Module google_api_proto::google::cloud::documentai::v1beta2::document

Nested message and enum types in Document.

Modules

label

Nested message and enum types in Label.

page

Nested message and enum types in Page.

page_anchor

Nested message and enum types in PageAnchor.

style

Nested message and enum types in Style.

text_anchor

Nested message and enum types in TextAnchor.

Structs

Entity

A phrase in the text that is a known entity type, such as a person, an organization, or location.

EntityRelation

Relationship between [Entities][google.cloud.documentai.v1beta2.Document.Entity].

Label

Label attaches schema information and/or other metadata to segments within a [Document][google.cloud.documentai.v1beta2.Document]. Multiple [Label][google.cloud.documentai.v1beta2.Document.Label]s on a single field can denote either different labels, different instances of the same label created at different times, or some combination of both.

Page

A page in a [Document][google.cloud.documentai.v1beta2.Document].

PageAnchor

Referencing elements in [Document.pages][google.cloud.documentai.v1beta2.Document.pages].

ShardInfo

For a large document, sharding may be performed to produce several document shards. Each document shard contains this field to detail which shard it is.

Style

Annotation for common text style attributes. This adheres to CSS conventions as much as possible.

TextAnchor

Text reference indexing into the [Document.text][google.cloud.documentai.v1beta2.Document.text].

Enums

Source

Original source document from the user.