Module google_api_proto::google::cloud::documentai::v1beta2::document::page

Nested message and enum types in Page.

Modules

image_quality_scores

Nested message and enum types in ImageQualityScores.

layout

Nested message and enum types in Layout.

table

Nested message and enum types in Table.

token

Nested message and enum types in Token.

Structs

Block

A block has a set of lines (collected into paragraphs) that have a common line-spacing and orientation.

DetectedBarcode

A detected barcode.

DetectedLanguage

Detected language for a structural component.

Dimension

Dimension for the page.

FormField

A form field detected on the page.

Image

Rendered image contents for this page.

ImageQualityScores

Image quality scores for the page image.

Layout

Visual element describing a layout unit on a page.

Line

A collection of tokens that a human would perceive as a line. Does not cross column boundaries, can be horizontal, vertical, etc.

Matrix

Representation for transformation matrix, intended to be compatible and used with OpenCV format for image manipulation.

Paragraph

A collection of lines that a human would perceive as a paragraph.

Symbol

A detected symbol.

Table

A table representation similar to HTML table structure.

Token

A detected token.

VisualElement

Detected non-text visual elements e.g. checkbox, signature etc. on the page.