Struct google_api_proto::google::cloud::documentai::v1beta3::BatchProcessRequest

pub struct BatchProcessRequest {
    pub name: String,
    pub input_configs: Vec<BatchInputConfig>,
    pub output_config: Option<BatchOutputConfig>,
    pub input_documents: Option<BatchDocumentsInputConfig>,
    pub document_output_config: Option<DocumentOutputConfig>,
    pub skip_human_review: bool,
    pub process_options: Option<ProcessOptions>,
}

Request message for [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments].

Fields
name: String

Required. The resource name of [Processor][google.cloud.documentai.v1beta3.Processor] or [ProcessorVersion][google.cloud.documentai.v1beta3.ProcessorVersion]. Format: projects/{project}/locations/{location}/processors/{processor}, or projects/{project}/locations/{location}/processors/{processor}/processorVersions/{processorVersion}

input_configs: Vec<BatchInputConfig>
๐Deprecated

The input config for each single document in the batch process.

output_config: Option<BatchOutputConfig>
๐Deprecated

The overall output config for batch process.

input_documents: Option<BatchDocumentsInputConfig>

The input documents for the [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments] method.

document_output_config: Option<DocumentOutputConfig>

The output configuration for the [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments] method.

skip_human_review: bool

Whether human review should be skipped for this request. Default to false.

process_options: Option<ProcessOptions>

Inference-time options for the process API