Module google_api_proto::google::cloud::domains::v1

Modules

contact_settings

Nested message and enum types in ContactSettings.

dns_settings

Nested message and enum types in DnsSettings.

domains_client

Generated client implementations.

management_settings

Nested message and enum types in ManagementSettings.

register_parameters

Nested message and enum types in RegisterParameters.

registration

Nested message and enum types in Registration.

Structs

AuthorizationCode

Defines an authorization code.

ConfigureContactSettingsRequest

Request for the ConfigureContactSettings method.

ConfigureDnsSettingsRequest

Request for the ConfigureDnsSettings method.

ConfigureManagementSettingsRequest

Request for the ConfigureManagementSettings method.

ContactSettings

Defines the contact information associated with a Registration.

DeleteRegistrationRequest

Request for the DeleteRegistration method.

DnsSettings

Defines the DNS configuration of a Registration, including name servers, DNSSEC, and glue records.

ExportRegistrationRequest

Request for the ExportRegistration method.

GetRegistrationRequest

Request for the GetRegistration method.

ListRegistrationsRequest

Request for the ListRegistrations method.

ListRegistrationsResponse

Response for the ListRegistrations method.

ManagementSettings

Defines renewal, billing, and transfer settings for a Registration.

OperationMetadata

Represents the metadata of the long-running operation. Output only.

RegisterDomainRequest

Request for the RegisterDomain method.

RegisterParameters

Parameters required to register a new domain.

Registration

The Registration resource facilitates managing and configuring domain name registrations.

ResetAuthorizationCodeRequest

Request for the ResetAuthorizationCode method.

RetrieveAuthorizationCodeRequest

Request for the RetrieveAuthorizationCode method.

RetrieveRegisterParametersRequest

Request for the RetrieveRegisterParameters method.

RetrieveRegisterParametersResponse

Response for the RetrieveRegisterParameters method.

RetrieveTransferParametersRequest

Request for the RetrieveTransferParameters method.

RetrieveTransferParametersResponse

Response for the RetrieveTransferParameters method.

SearchDomainsRequest

Request for the SearchDomains method.

SearchDomainsResponse

Response for the SearchDomains method.

TransferDomainRequest

Request for the TransferDomain method.

TransferParameters

Parameters required to transfer a domain from another registrar.

UpdateRegistrationRequest

Request for the UpdateRegistration method.

Enums

ContactNotice

Notices related to contact information.

ContactPrivacy

Defines a set of possible contact privacy settings for a Registration.

DomainNotice

Notices about special properties of certain domains.

TransferLockState

Possible states of a Registration’s transfer lock.