Struct google_api_proto::google::cloud::domains::v1beta1::domains_client::DomainsClient

pub struct DomainsClient<T> { /* private fields */ }

The Cloud Domains API enables management and configuration of domain names.

Implementations

impl<T> DomainsClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> DomainsClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn search_domains( &mut self, request: impl IntoRequest<SearchDomainsRequest>, ) -> Result<Response<SearchDomainsResponse>, Status>

Searches for available domain names similar to the provided query.

Availability results from this method are approximate; call RetrieveRegisterParameters on a domain before registering to confirm availability.

pub async fn retrieve_register_parameters( &mut self, request: impl IntoRequest<RetrieveRegisterParametersRequest>, ) -> Result<Response<RetrieveRegisterParametersResponse>, Status>

Gets parameters needed to register a new domain name, including price and up-to-date availability. Use the returned values to call RegisterDomain.

pub async fn register_domain( &mut self, request: impl IntoRequest<RegisterDomainRequest>, ) -> Result<Response<Operation>, Status>

Registers a new domain name and creates a corresponding Registration resource.

Call RetrieveRegisterParameters first to check availability of the domain name and determine parameters like price that are needed to build a call to this method.

A successful call creates a Registration resource in state REGISTRATION_PENDING, which resolves to ACTIVE within 1-2 minutes, indicating that the domain was successfully registered. If the resource ends up in state REGISTRATION_FAILED, it indicates that the domain was not registered successfully, and you can safely delete the resource and retry registration.

pub async fn retrieve_transfer_parameters( &mut self, request: impl IntoRequest<RetrieveTransferParametersRequest>, ) -> Result<Response<RetrieveTransferParametersResponse>, Status>

Gets parameters needed to transfer a domain name from another registrar to Cloud Domains. For domains managed by Google Domains, transferring to Cloud Domains is not supported.

Use the returned values to call TransferDomain.

pub async fn transfer_domain( &mut self, request: impl IntoRequest<TransferDomainRequest>, ) -> Result<Response<Operation>, Status>

Transfers a domain name from another registrar to Cloud Domains. For domains managed by Google Domains, transferring to Cloud Domains is not supported.

Before calling this method, go to the domain’s current registrar to unlock the domain for transfer and retrieve the domain’s transfer authorization code. Then call RetrieveTransferParameters to confirm that the domain is unlocked and to get values needed to build a call to this method.

A successful call creates a Registration resource in state TRANSFER_PENDING. It can take several days to complete the transfer process. The registrant can often speed up this process by approving the transfer through the current registrar, either by clicking a link in an email from the registrar or by visiting the registrar’s website.

A few minutes after transfer approval, the resource transitions to state ACTIVE, indicating that the transfer was successful. If the transfer is rejected or the request expires without being approved, the resource can end up in state TRANSFER_FAILED. If transfer fails, you can safely delete the resource and retry the transfer.

pub async fn list_registrations( &mut self, request: impl IntoRequest<ListRegistrationsRequest>, ) -> Result<Response<ListRegistrationsResponse>, Status>

Lists the Registration resources in a project.

pub async fn get_registration( &mut self, request: impl IntoRequest<GetRegistrationRequest>, ) -> Result<Response<Registration>, Status>

Gets the details of a Registration resource.

pub async fn update_registration( &mut self, request: impl IntoRequest<UpdateRegistrationRequest>, ) -> Result<Response<Operation>, Status>

Updates select fields of a Registration resource, notably labels. To update other fields, use the appropriate custom update method:

To update management settings, see ConfigureManagementSettings

To update DNS configuration, see ConfigureDnsSettings

To update contact information, see ConfigureContactSettings

pub async fn configure_management_settings( &mut self, request: impl IntoRequest<ConfigureManagementSettingsRequest>, ) -> Result<Response<Operation>, Status>

Updates a Registration’s management settings.

pub async fn configure_dns_settings( &mut self, request: impl IntoRequest<ConfigureDnsSettingsRequest>, ) -> Result<Response<Operation>, Status>

Updates a Registration’s DNS settings.

pub async fn configure_contact_settings( &mut self, request: impl IntoRequest<ConfigureContactSettingsRequest>, ) -> Result<Response<Operation>, Status>

Updates a Registration’s contact settings. Some changes require confirmation by the domain’s registrant contact .

pub async fn export_registration( &mut self, request: impl IntoRequest<ExportRegistrationRequest>, ) -> Result<Response<Operation>, Status>

Exports a Registration resource, such that it is no longer managed by Cloud Domains.

When an active domain is successfully exported, you can continue to use the domain in Google Domains until it expires. The calling user becomes the domain’s sole owner in Google Domains, and permissions for the domain are subsequently managed there. The domain does not renew automatically unless the new owner sets up billing in Google Domains.

pub async fn delete_registration( &mut self, request: impl IntoRequest<DeleteRegistrationRequest>, ) -> Result<Response<Operation>, Status>

Deletes a Registration resource.

This method works on any Registration resource using Subscription or Commitment billing, provided that the resource was created at least 1 day in the past.

For Registration resources using Monthly billing, this method works if:

state is EXPORTED with expire_time in the past

state is REGISTRATION_FAILED

state is TRANSFER_FAILED

When an active registration is successfully deleted, you can continue to use the domain in Google Domains until it expires. The calling user becomes the domain’s sole owner in Google Domains, and permissions for the domain are subsequently managed there. The domain does not renew automatically unless the new owner sets up billing in Google Domains.

pub async fn retrieve_authorization_code( &mut self, request: impl IntoRequest<RetrieveAuthorizationCodeRequest>, ) -> Result<Response<AuthorizationCode>, Status>

Gets the authorization code of the Registration for the purpose of transferring the domain to another registrar.

You can call this method only after 60 days have elapsed since the initial domain registration.

pub async fn reset_authorization_code( &mut self, request: impl IntoRequest<ResetAuthorizationCodeRequest>, ) -> Result<Response<AuthorizationCode>, Status>

Resets the authorization code of the Registration to a new random string.

You can call this method only after 60 days have elapsed since the initial domain registration.