Module google_api_proto::google::cloud::edgecontainer::v1::cluster

Nested message and enum types in Cluster.

Modules

control_plane

Nested message and enum types in ControlPlane.

maintenance_event

Nested message and enum types in MaintenanceEvent.

system_addons_config

Nested message and enum types in SystemAddonsConfig.

Structs

ControlPlane

Configuration of the cluster control plane.

ControlPlaneEncryption

Configuration for Customer-managed KMS key support for remote control plane cluster disk encryption.

MaintenanceEvent

A Maintenance Event is an operation that could cause temporary disruptions to the cluster workloads, including Google-driven or user-initiated cluster upgrades, user-initiated cluster configuration changes that require restarting nodes, etc.

SurvivabilityConfig

Configuration of the cluster survivability, e.g., for the case when network connectivity is lost.

SystemAddonsConfig

Config that customers are allowed to define for GDCE system add-ons.

Enums

ReleaseChannel

The release channel a cluster is subscribed to.

Status

Indicates the status of the cluster.