Struct google_api_proto::google::cloud::edgecontainer::v1::Cluster

pub struct Cluster {

    pub name: String,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub labels: BTreeMap<String, String>,
    pub fleet: Option<Fleet>,
    pub networking: Option<ClusterNetworking>,
    pub authorization: Option<Authorization>,
    pub default_max_pods_per_node: i32,
    pub endpoint: String,
    pub port: i32,
    pub cluster_ca_certificate: String,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub control_plane_version: String,
    pub node_version: String,
    pub control_plane: Option<ControlPlane>,
    pub system_addons_config: Option<SystemAddonsConfig>,
    pub external_load_balancer_ipv4_address_pools: Vec<String>,
    pub control_plane_encryption: Option<ControlPlaneEncryption>,
    pub status: i32,
    pub maintenance_events: Vec<MaintenanceEvent>,
    pub target_version: String,
    pub release_channel: i32,
    pub survivability_config: Option<SurvivabilityConfig>,
    pub external_load_balancer_ipv6_address_pools: Vec<String>,

}

A Google Distributed Cloud Edge Kubernetes cluster.

Fields
name: String

Required. The resource name of the cluster.

create_time: Option<Timestamp>

Output only. The time when the cluster was created.

update_time: Option<Timestamp>

Output only. The time when the cluster was last updated.

labels: BTreeMap<String, String>

Labels associated with this resource.

fleet: Option<Fleet>

Required. Fleet configuration.

networking: Option<ClusterNetworking>

Required. Cluster-wide networking configuration.

authorization: Option<Authorization>

Required. Immutable. RBAC policy that will be applied and managed by GEC.

default_max_pods_per_node: i32

Optional. The default maximum number of pods per node used if a maximum value is not specified explicitly for a node pool in this cluster. If unspecified, the Kubernetes default value will be used.

endpoint: String

Output only. The IP address of the Kubernetes API server.

port: i32

Output only. The port number of the Kubernetes API server.

cluster_ca_certificate: String

Output only. The PEM-encoded public certificate of the cluster’s CA.

maintenance_policy: Option<MaintenancePolicy>

Optional. Cluster-wide maintenance policy configuration.

control_plane_version: String

Output only. The control plane release version

node_version: String

Output only. The lowest release version among all worker nodes. This field can be empty if the cluster does not have any worker nodes.

control_plane: Option<ControlPlane>

Optional. The configuration of the cluster control plane.

system_addons_config: Option<SystemAddonsConfig>

Optional. The configuration of the system add-ons.

external_load_balancer_ipv4_address_pools: Vec<String>

Optional. IPv4 address pools for cluster data plane external load balancing.

control_plane_encryption: Option<ControlPlaneEncryption>

Optional. Remote control plane disk encryption options. This field is only used when enabling CMEK support.

status: i32

Output only. The current status of the cluster.

maintenance_events: Vec<MaintenanceEvent>

Output only. All the maintenance events scheduled for the cluster, including the ones ongoing, planned for the future and done in the past (up to 90 days).

target_version: String

Optional. The target cluster version. For example: “1.5.0”.

release_channel: i32

Optional. The release channel a cluster is subscribed to.

survivability_config: Option<SurvivabilityConfig>

Optional. Configuration of the cluster survivability, e.g., for the case when network connectivity is lost. Note: This only applies to local control plane clusters.

external_load_balancer_ipv6_address_pools: Vec<String>

Optional. IPv6 address pools for cluster data plane external load balancing.

Implementations

impl Cluster

pub fn status(&self) -> Status

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: Status)

Sets status to the provided enum value.

pub fn release_channel(&self) -> ReleaseChannel

Returns the enum value of release_channel, or the default if the field is set to an invalid enum value.

pub fn set_release_channel(&mut self, value: ReleaseChannel)

Sets release_channel to the provided enum value.