Module google_api_proto::google::cloud::functions::v2beta

Modules

build_config

Nested message and enum types in BuildConfig.

event_trigger

Nested message and enum types in EventTrigger.

function

Nested message and enum types in Function.

function_service_client

Generated client implementations.

list_runtimes_response

Nested message and enum types in ListRuntimesResponse.

repo_source

Nested message and enum types in RepoSource.

secret_volume

Nested message and enum types in SecretVolume.

service_config

Nested message and enum types in ServiceConfig.

source

Nested message and enum types in Source.

stage

Nested message and enum types in Stage.

state_message

Nested message and enum types in StateMessage.

Structs

AutomaticUpdatePolicy

Security patches are applied automatically to the runtime without requiring the function to be redeployed.

BuildConfig

Describes the Build step of the function that builds a container from the given source.

CreateFunctionRequest

Request for the CreateFunction method.

DeleteFunctionRequest

Request for the DeleteFunction method.

EventFilter

Filters events based on exact matches on the CloudEvents attributes.

EventTrigger

Describes EventTrigger, used to request events to be sent from another service.

Function

Describes a Cloud Function that contains user computation executed in response to an event. It encapsulates function and trigger configurations.

GenerateDownloadUrlRequest

Request of GenerateDownloadUrl method.

GenerateDownloadUrlResponse

Response of GenerateDownloadUrl method.

GenerateUploadUrlRequest

Request of GenerateSourceUploadUrl method.

GenerateUploadUrlResponse

Response of GenerateSourceUploadUrl method.

GetFunctionRequest

Request for the GetFunction method.

ListFunctionsRequest

Request for the ListFunctions method.

ListFunctionsResponse

Response for the ListFunctions method.

ListRuntimesRequest

Request for the ListRuntimes method.

ListRuntimesResponse

Response for the ListRuntimes method.

LocationMetadata

Extra GCF specific location information.

OnDeployUpdatePolicy

Security patches are only applied when a function is redeployed.

OperationMetadata

Represents the metadata of the long-running operation.

RepoSource

Location of the source in a Google Cloud Source Repository.

SecretEnvVar

Configuration for a secret environment variable. It has the information necessary to fetch the secret value from secret manager and expose it as an environment variable.

SecretVolume

Configuration for a secret volume. It has the information necessary to fetch the secret value from secret manager and make it available as files mounted at the requested paths within the application container.

ServiceConfig

Describes the Service being deployed. Currently Supported : Cloud Run (fully managed).

Source

The location of the function source code.

SourceProvenance

Provenance of the source. Ways to find the original source, or verify that some source was used for this build.

Stage

Each Stage of the deployment process

StateMessage

Informational messages about the state of the Cloud Function or Operation.

StorageSource

Location of the source in an archive file in Google Cloud Storage.

UpdateFunctionRequest

Request for the UpdateFunction method.

Enums

Environment

The environment the function is hosted on.

OperationType

The type of the long running operation.