Module google_api_proto::google::cloud::gdchardwaremanagement::v1alpha

Modules

gdc_hardware_management_client

Generated client implementations.

hardware

Nested message and enum types in Hardware.

hardware_group

Nested message and enum types in HardwareGroup.

hardware_installation_info

Nested message and enum types in HardwareInstallationInfo.

hardware_physical_info

Nested message and enum types in HardwarePhysicalInfo.

order

Nested message and enum types in Order.

signal_zone_state_request

Nested message and enum types in SignalZoneStateRequest.

sku

Nested message and enum types in Sku.

zone

Nested message and enum types in Zone.

Structs

ChangeLogEntry

A log entry of a change made to an order.

Comment

A comment on an order.

Contact

Contact details of a point of contact.

CreateCommentRequest

A request to create a comment.

CreateHardwareGroupRequest

A request to create a hardware group.

CreateHardwareRequest

A request to create hardware.

CreateOrderRequest

A request to create an order.

CreateSiteRequest

A request to create a site.

CreateZoneRequest

A request to create a zone.

DeleteHardwareGroupRequest

A request to delete a hardware group.

DeleteHardwareRequest

A request to delete hardware.

DeleteOrderRequest

A request to delete an order.

DeleteZoneRequest

A request to delete a zone.

Dimensions

Represents the dimensions of an object.

GetChangeLogEntryRequest

A request to get a change log entry.

GetCommentRequest

A request to get a comment.

GetHardwareGroupRequest

A request to get a hardware group.

GetHardwareRequest

A request to get hardware.

GetOrderRequest

A request to get an order.

GetSiteRequest

A request to get a site.

GetSkuRequest

A request to get an SKU.

GetZoneRequest

A request to get a zone.

Hardware

An instance of hardware installed at a site.

HardwareConfig

Configuration for GDC hardware.

HardwareGroup

A group of hardware that is part of the same order, has the same SKU, and is delivered to the same site.

HardwareInstallationInfo

Information for installation of a Hardware.

HardwareLocation

Represents the location of one or many hardware.

HardwarePhysicalInfo

Physical properties of a hardware.

ListChangeLogEntriesRequest

A request to list change log entries.

ListChangeLogEntriesResponse

A list of change log entries.

ListCommentsRequest

A request to list comments.

ListCommentsResponse

A request to list comments.

ListHardwareGroupsRequest

A request to list hardware groups.

ListHardwareGroupsResponse

A list of hardware groups.

ListHardwareRequest

A request to list hardware.

ListHardwareResponse

A list of hardware.

ListOrdersRequest

A request to list orders.

ListOrdersResponse

A list of orders.

ListSitesRequest

A request to list sites.

ListSitesResponse

A list of sites.

ListSkusRequest

A request to list SKUs.

ListSkusResponse

A list of SKUs.

ListZonesRequest

A request to list zones.

ListZonesResponse

A list of zones.

OperationMetadata

Represents the metadata of a long-running operation.

Order

An order for GDC hardware.

OrganizationContact

Contact information of the customer organization.

RackSpace

Represents contiguous space in a rack.

SignalZoneStateRequest

A request to signal the state of a zone.

Site

A physical site where hardware will be installed.

Sku

A stock keeping unit (SKU) of GDC hardware.

SkuConfig

Configuration for a SKU.

SkuInstance

A specific instance of the SKU.

SubmitOrderRequest

A request to submit an order.

Subnet

Represents a subnet.

TimePeriod

Represents a time period in a week.

UpdateHardwareGroupRequest

A request to update a hardware group.

UpdateHardwareRequest

A request to update hardware.

UpdateOrderRequest

A request to update an order.

UpdateSiteRequest

A request to update a site.

UpdateZoneRequest

A request to update a zone.

Zone

A zone holding a set of hardware.

ZoneNetworkConfig

Networking configuration for a zone.

Enums

PowerSupply

The power supply options.