Struct google_api_proto::google::cloud::gdchardwaremanagement::v1alpha::gdc_hardware_management_client::GdcHardwareManagementClient

pub struct GdcHardwareManagementClient<T> { /* private fields */ }

The GDC Hardware Management service.

Implementations

impl<T> GdcHardwareManagementClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> GdcHardwareManagementClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_orders( &mut self, request: impl IntoRequest<ListOrdersRequest>, ) -> Result<Response<ListOrdersResponse>, Status>

Lists orders in a given project and location.

pub async fn get_order( &mut self, request: impl IntoRequest<GetOrderRequest>, ) -> Result<Response<Order>, Status>

Gets details of an order.

pub async fn create_order( &mut self, request: impl IntoRequest<CreateOrderRequest>, ) -> Result<Response<Operation>, Status>

Creates a new order in a given project and location.

pub async fn update_order( &mut self, request: impl IntoRequest<UpdateOrderRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of an order.

pub async fn delete_order( &mut self, request: impl IntoRequest<DeleteOrderRequest>, ) -> Result<Response<Operation>, Status>

Deletes an order.

pub async fn submit_order( &mut self, request: impl IntoRequest<SubmitOrderRequest>, ) -> Result<Response<Operation>, Status>

Submits an order.

pub async fn list_sites( &mut self, request: impl IntoRequest<ListSitesRequest>, ) -> Result<Response<ListSitesResponse>, Status>

Lists sites in a given project and location.

pub async fn get_site( &mut self, request: impl IntoRequest<GetSiteRequest>, ) -> Result<Response<Site>, Status>

Gets details of a site.

pub async fn create_site( &mut self, request: impl IntoRequest<CreateSiteRequest>, ) -> Result<Response<Operation>, Status>

Creates a new site in a given project and location.

pub async fn update_site( &mut self, request: impl IntoRequest<UpdateSiteRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a site.

pub async fn list_hardware_groups( &mut self, request: impl IntoRequest<ListHardwareGroupsRequest>, ) -> Result<Response<ListHardwareGroupsResponse>, Status>

Lists hardware groups in a given order.

pub async fn get_hardware_group( &mut self, request: impl IntoRequest<GetHardwareGroupRequest>, ) -> Result<Response<HardwareGroup>, Status>

Gets details of a hardware group.

pub async fn create_hardware_group( &mut self, request: impl IntoRequest<CreateHardwareGroupRequest>, ) -> Result<Response<Operation>, Status>

Creates a new hardware group in a given order.

pub async fn update_hardware_group( &mut self, request: impl IntoRequest<UpdateHardwareGroupRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a hardware group.

pub async fn delete_hardware_group( &mut self, request: impl IntoRequest<DeleteHardwareGroupRequest>, ) -> Result<Response<Operation>, Status>

Deletes a hardware group.

pub async fn list_hardware( &mut self, request: impl IntoRequest<ListHardwareRequest>, ) -> Result<Response<ListHardwareResponse>, Status>

Lists hardware in a given project and location.

pub async fn get_hardware( &mut self, request: impl IntoRequest<GetHardwareRequest>, ) -> Result<Response<Hardware>, Status>

Gets hardware details.

pub async fn create_hardware( &mut self, request: impl IntoRequest<CreateHardwareRequest>, ) -> Result<Response<Operation>, Status>

Creates new hardware in a given project and location.

pub async fn update_hardware( &mut self, request: impl IntoRequest<UpdateHardwareRequest>, ) -> Result<Response<Operation>, Status>

Updates hardware parameters.

pub async fn delete_hardware( &mut self, request: impl IntoRequest<DeleteHardwareRequest>, ) -> Result<Response<Operation>, Status>

Deletes hardware.

pub async fn list_comments( &mut self, request: impl IntoRequest<ListCommentsRequest>, ) -> Result<Response<ListCommentsResponse>, Status>

Lists the comments on an order.

pub async fn get_comment( &mut self, request: impl IntoRequest<GetCommentRequest>, ) -> Result<Response<Comment>, Status>

Gets the content of a comment.

pub async fn create_comment( &mut self, request: impl IntoRequest<CreateCommentRequest>, ) -> Result<Response<Operation>, Status>

Creates a new comment on an order.

pub async fn list_change_log_entries( &mut self, request: impl IntoRequest<ListChangeLogEntriesRequest>, ) -> Result<Response<ListChangeLogEntriesResponse>, Status>

Lists the changes made to an order.

pub async fn get_change_log_entry( &mut self, request: impl IntoRequest<GetChangeLogEntryRequest>, ) -> Result<Response<ChangeLogEntry>, Status>

Gets details of a change to an order.

pub async fn list_skus( &mut self, request: impl IntoRequest<ListSkusRequest>, ) -> Result<Response<ListSkusResponse>, Status>

Lists SKUs for a given project and location.

pub async fn get_sku( &mut self, request: impl IntoRequest<GetSkuRequest>, ) -> Result<Response<Sku>, Status>

Gets details of an SKU.

pub async fn list_zones( &mut self, request: impl IntoRequest<ListZonesRequest>, ) -> Result<Response<ListZonesResponse>, Status>

Lists zones in a given project and location.

pub async fn get_zone( &mut self, request: impl IntoRequest<GetZoneRequest>, ) -> Result<Response<Zone>, Status>

Gets details of a zone.

pub async fn create_zone( &mut self, request: impl IntoRequest<CreateZoneRequest>, ) -> Result<Response<Operation>, Status>

Creates a new zone in a given project and location.

pub async fn update_zone( &mut self, request: impl IntoRequest<UpdateZoneRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a zone.

pub async fn delete_zone( &mut self, request: impl IntoRequest<DeleteZoneRequest>, ) -> Result<Response<Operation>, Status>

Deletes a zone.

pub async fn signal_zone_state( &mut self, request: impl IntoRequest<SignalZoneStateRequest>, ) -> Result<Response<Operation>, Status>

Signals the state of a zone.