Module google_api_proto::google::cloud::gkebackup::v1::restore_config

Nested message and enum types in RestoreConfig.

Modules

restore_order

Nested message and enum types in RestoreOrder.

transformation_rule_action

Nested message and enum types in TransformationRuleAction.

volume_data_restore_policy_binding

Nested message and enum types in VolumeDataRestorePolicyBinding.

Structs

ClusterResourceRestoreScope

Defines the scope of cluster-scoped resources to restore.

GroupKind

This is a direct map to the Kubernetes GroupKind type GroupKind and is used for identifying specific “types” of resources to restore.

ResourceFilter

ResourceFilter specifies matching criteria to limit the scope of a change to a specific set of kubernetes resources that are selected for restoration from a backup.

RestoreOrder

Allows customers to specify dependencies between resources that Backup for GKE can use to compute a resasonable restore order.

SubstitutionRule

A transformation rule to be applied against Kubernetes resources as they are selected for restoration from a Backup. A rule contains both filtering logic (which resources are subject to substitution) and substitution logic.

TransformationRule

A transformation rule to be applied against Kubernetes resources as they are selected for restoration from a Backup. A rule contains both filtering logic (which resources are subject to transform) and transformation logic.

TransformationRuleAction

TransformationRuleAction defines a TransformationRule action based on the JSON Patch RFC (https://www.rfc-editor.org/rfc/rfc6902)

VolumeDataRestorePolicyBinding

Binds resources in the scope to the given VolumeDataRestorePolicy.

Enums

ClusterResourceConflictPolicy

Defines the behavior for handling the situation where cluster-scoped resources being restored already exist in the target cluster.

NamespacedResourceRestoreMode

Defines the behavior for handling the situation where sets of namespaced resources being restored already exist in the target cluster.

NamespacedResourceRestoreScope

Specifies the namespaced resources to restore from the Backup. Only one of the entries may be specified. If not specified, NO namespaced resources will be restored.

VolumeDataRestorePolicy

Defines how volume data should be restored.