Struct google_api_proto::google::cloud::gkebackup::v1::RestoreConfig

pub struct RestoreConfig {
    pub volume_data_restore_policy: i32,
    pub cluster_resource_conflict_policy: i32,
    pub namespaced_resource_restore_mode: i32,
    pub cluster_resource_restore_scope: Option<ClusterResourceRestoreScope>,
    pub substitution_rules: Vec<SubstitutionRule>,
    pub transformation_rules: Vec<TransformationRule>,
    pub volume_data_restore_policy_bindings: Vec<VolumeDataRestorePolicyBinding>,
    pub restore_order: Option<RestoreOrder>,
    pub namespaced_resource_restore_scope: Option<NamespacedResourceRestoreScope>,
}

Configuration of a restore.

Fields
volume_data_restore_policy: i32

Optional. Specifies the mechanism to be used to restore volume data. Default: VOLUME_DATA_RESTORE_POLICY_UNSPECIFIED (will be treated as NO_VOLUME_DATA_RESTORATION).

cluster_resource_conflict_policy: i32

Optional. Defines the behavior for handling the situation where cluster-scoped resources being restored already exist in the target cluster. This MUST be set to a value other than CLUSTER_RESOURCE_CONFLICT_POLICY_UNSPECIFIED if [cluster_resource_restore_scope][google.cloud.gkebackup.v1.RestoreConfig.cluster_resource_restore_scope] is not empty.

namespaced_resource_restore_mode: i32

Optional. Defines the behavior for handling the situation where sets of namespaced resources being restored already exist in the target cluster. This MUST be set to a value other than NAMESPACED_RESOURCE_RESTORE_MODE_UNSPECIFIED.

cluster_resource_restore_scope: Option<ClusterResourceRestoreScope>

Optional. Identifies the cluster-scoped resources to restore from the Backup. Not specifying it means NO cluster resource will be restored.

substitution_rules: Vec<SubstitutionRule>

Optional. A list of transformation rules to be applied against Kubernetes resources as they are selected for restoration from a Backup. Rules are executed in order defined - this order matters, as changes made by a rule may impact the filtering logic of subsequent rules. An empty list means no substitution will occur.

transformation_rules: Vec<TransformationRule>

Optional. A list of transformation rules to be applied against Kubernetes resources as they are selected for restoration from a Backup. Rules are executed in order defined - this order matters, as changes made by a rule may impact the filtering logic of subsequent rules. An empty list means no transformation will occur.

volume_data_restore_policy_bindings: Vec<VolumeDataRestorePolicyBinding>

Optional. A table that binds volumes by their scope to a restore policy. Bindings must have a unique scope. Any volumes not scoped in the bindings are subject to the policy defined in volume_data_restore_policy.

restore_order: Option<RestoreOrder>

Optional. RestoreOrder contains custom ordering to use on a Restore.

namespaced_resource_restore_scope: Option<NamespacedResourceRestoreScope>

Specifies the namespaced resources to restore from the Backup. Only one of the entries may be specified. If not specified, NO namespaced resources will be restored.

Note: Resources will never be restored into managed namespaces such as kube-system, kube-public, or kube-node-lease. These namespaces are silently skipped when [all_namespaces][google.cloud.gkebackup.v1.RestoreConfig.all_namespaces] is selected. Listing them explicitly will result in an error.

Implementations

impl RestoreConfig

pub fn volume_data_restore_policy(&self) -> VolumeDataRestorePolicy

Returns the enum value of volume_data_restore_policy, or the default if the field is set to an invalid enum value.

pub fn set_volume_data_restore_policy(&mut self, value: VolumeDataRestorePolicy)

Sets volume_data_restore_policy to the provided enum value.

pub fn cluster_resource_conflict_policy(&self) -> ClusterResourceConflictPolicy

Returns the enum value of cluster_resource_conflict_policy, or the default if the field is set to an invalid enum value.

pub fn set_cluster_resource_conflict_policy( &mut self, value: ClusterResourceConflictPolicy, )

Sets cluster_resource_conflict_policy to the provided enum value.

pub fn namespaced_resource_restore_mode(&self) -> NamespacedResourceRestoreMode

Returns the enum value of namespaced_resource_restore_mode, or the default if the field is set to an invalid enum value.

pub fn set_namespaced_resource_restore_mode( &mut self, value: NamespacedResourceRestoreMode, )

Sets namespaced_resource_restore_mode to the provided enum value.