Struct google_api_proto::google::cloud::gkemulticloud::v1::aws_clusters_client::AwsClustersClient

pub struct AwsClustersClient<T> { /* private fields */ }

The AwsClusters API provides a single centrally managed service to create and manage Anthos clusters that run on AWS infrastructure.

Implementations

impl<T> AwsClustersClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> AwsClustersClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_aws_cluster( &mut self, request: impl IntoRequest<CreateAwsClusterRequest>, ) -> Result<Response<Operation>, Status>

Creates a new [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster] resource on a given Google Cloud Platform project and region.

If successful, the response contains a newly created [Operation][google.longrunning.Operation] resource that can be described to track the status of the operation.

pub async fn update_aws_cluster( &mut self, request: impl IntoRequest<UpdateAwsClusterRequest>, ) -> Result<Response<Operation>, Status>

Updates an [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster].

pub async fn get_aws_cluster( &mut self, request: impl IntoRequest<GetAwsClusterRequest>, ) -> Result<Response<AwsCluster>, Status>

Describes a specific [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster] resource.

pub async fn list_aws_clusters( &mut self, request: impl IntoRequest<ListAwsClustersRequest>, ) -> Result<Response<ListAwsClustersResponse>, Status>

Lists all [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster] resources on a given Google Cloud project and region.

pub async fn delete_aws_cluster( &mut self, request: impl IntoRequest<DeleteAwsClusterRequest>, ) -> Result<Response<Operation>, Status>

Deletes a specific [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster] resource.

Fails if the cluster has one or more associated [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool] resources.

If successful, the response contains a newly created [Operation][google.longrunning.Operation] resource that can be described to track the status of the operation.

pub async fn generate_aws_cluster_agent_token( &mut self, request: impl IntoRequest<GenerateAwsClusterAgentTokenRequest>, ) -> Result<Response<GenerateAwsClusterAgentTokenResponse>, Status>

Generates an access token for a cluster agent.

pub async fn generate_aws_access_token( &mut self, request: impl IntoRequest<GenerateAwsAccessTokenRequest>, ) -> Result<Response<GenerateAwsAccessTokenResponse>, Status>

Generates a short-lived access token to authenticate to a given [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster] resource.

pub async fn create_aws_node_pool( &mut self, request: impl IntoRequest<CreateAwsNodePoolRequest>, ) -> Result<Response<Operation>, Status>

Creates a new [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool], attached to a given [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster].

If successful, the response contains a newly created [Operation][google.longrunning.Operation] resource that can be described to track the status of the operation.

pub async fn update_aws_node_pool( &mut self, request: impl IntoRequest<UpdateAwsNodePoolRequest>, ) -> Result<Response<Operation>, Status>

Updates an [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool].

pub async fn rollback_aws_node_pool_update( &mut self, request: impl IntoRequest<RollbackAwsNodePoolUpdateRequest>, ) -> Result<Response<Operation>, Status>

Rolls back a previously aborted or failed [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool] update request. Makes no changes if the last update request successfully finished. If an update request is in progress, you cannot rollback the update. You must first cancel or let it finish unsuccessfully before you can rollback.

pub async fn get_aws_node_pool( &mut self, request: impl IntoRequest<GetAwsNodePoolRequest>, ) -> Result<Response<AwsNodePool>, Status>

Describes a specific [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool] resource.

pub async fn list_aws_node_pools( &mut self, request: impl IntoRequest<ListAwsNodePoolsRequest>, ) -> Result<Response<ListAwsNodePoolsResponse>, Status>

Lists all [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool] resources on a given [AwsCluster][google.cloud.gkemulticloud.v1.AwsCluster].

pub async fn delete_aws_node_pool( &mut self, request: impl IntoRequest<DeleteAwsNodePoolRequest>, ) -> Result<Response<Operation>, Status>

Deletes a specific [AwsNodePool][google.cloud.gkemulticloud.v1.AwsNodePool] resource.

If successful, the response contains a newly created [Operation][google.longrunning.Operation] resource that can be described to track the status of the operation.

pub async fn get_aws_open_id_config( &mut self, request: impl IntoRequest<GetAwsOpenIdConfigRequest>, ) -> Result<Response<AwsOpenIdConfig>, Status>

Gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

pub async fn get_aws_json_web_keys( &mut self, request: impl IntoRequest<GetAwsJsonWebKeysRequest>, ) -> Result<Response<AwsJsonWebKeys>, Status>

Gets the public component of the cluster signing keys in JSON Web Key format.

pub async fn get_aws_server_config( &mut self, request: impl IntoRequest<GetAwsServerConfigRequest>, ) -> Result<Response<AwsServerConfig>, Status>

Returns information, such as supported AWS regions and Kubernetes versions, on a given Google Cloud location.