Module google_api_proto::google::cloud::integrations::v1alpha

Modules

cloud_logging_details

Nested message and enum types in CloudLoggingDetails.

execution_info

Nested message and enum types in ExecutionInfo.

failure_policy

Nested message and enum types in FailurePolicy.

integration_execution_details

Nested message and enum types in IntegrationExecutionDetails.

integration_execution_snapshot

Nested message and enum types in IntegrationExecutionSnapshot.

success_policy

Nested message and enum types in SuccessPolicy.

task_config

Nested message and enum types in TaskConfig.

task_execution_details

Nested message and enum types in TaskExecutionDetails.

value_type

Nested message and enum types in ValueType.

Structs

AttemptStats

Status for the execution attempt.

BooleanParameterArray

This message only contains a field of boolean array.

CloudLoggingDetails

Cloud Logging details for execution info

ConditionResult

Contains the combined condition calculation results.

Coordinate

Configuration detail of coordinate, it used for UI

DoubleParameterArray

This message only contains a field of double number array.

ErrorDetail

An error, warning, or information message associated with an integration.

EventParameter

This message is used for processing and persisting (when applicable) key value pair parameters for each event in the event bus. Next available id: 4

ExecutionInfo

Log entry to log execution info for the monitored resource integrations.googleapis.com/IntegrationVersion.

FailurePolicy

Policy that defines the task retry logic and failure type. If no FailurePolicy is defined for a task, all its dependent tasks will not be executed (i.e, a retry_strategy of NONE will be applied).

IntParameterArray

This message only contains a field of integer array.

IntegrationExecutionDetails

Contains the details of the execution info: this includes the tasks execution details plus the integration execution statistics.

IntegrationExecutionSnapshot

Contains the snapshot of the integration execution for a given checkpoint.

NextTask

The task that is next in line to be executed, if the condition specified evaluated to true.

StringParameterArray

This message only contains a field of string array.

SuccessPolicy

Policy that dictates the behavior for the task after it completes successfully.

TaskConfig

The task configuration details. This is not the implementation of Task. There might be multiple TaskConfigs for the same Task.

TaskExecutionDetails

Contains the details of the execution of this task.

ValueType

The type of the parameter.

Enums

ExecutionType

Specifies whether this execution info corresponds to actual integration or test case.

IntegrationState

Indicates the status of the integration.

JsonValidationOption

Options for how to validate json schemas.

Product

Enum Product.