Struct google_api_proto::google::cloud::integrations::v1alpha::ExecutionInfo

pub struct ExecutionInfo {

    pub integration: String,
    pub project_id: String,
    pub trigger_id: String,
    pub request_params: BTreeMap<String, EventParameter>,
    pub response_params: BTreeMap<String, EventParameter>,
    pub errors: Vec<ErrorDetail>,
    pub task_configs: Vec<TaskConfig>,
    pub integration_version_number: String,
    pub execution_id: String,
    pub integration_version_state: i32,
    pub enable_database_persistence: bool,
    pub cloud_logging_details: Option<CloudLoggingDetails>,
    pub integration_execution_details: Option<IntegrationExecutionDetails>,
    pub execution_type: i32,
    pub execution_method: i32,
    pub integration_snapshot_number: i64,

}

Log entry to log execution info for the monitored resource integrations.googleapis.com/IntegrationVersion.

Fields
integration: String

Name of the integration.

project_id: String

The customer’s project number.

trigger_id: String

The trigger id of the integration trigger config. If both trigger_id and client_id is present, the integration is executed from the start tasks provided by the matching trigger config otherwise it is executed from the default start tasks.

request_params: BTreeMap<String, EventParameter>

Execution parameters come in as part of the request.

response_params: BTreeMap<String, EventParameter>

Execution parameters come out as part of the response.

errors: Vec<ErrorDetail>

Errors, warnings, and informationals associated with the workflow/task. The order in which the errors were added by the workflow/task is maintained.

task_configs: Vec<TaskConfig>

The configuration details for a task.

integration_version_number: String

Pointer to the active version it is executing.

execution_id: String

Auto-generated primary key.

integration_version_state: i32

Output only. State of the integration version

enable_database_persistence: bool

Database persistence policy for execution info

cloud_logging_details: Option<CloudLoggingDetails>

Cloud Logging details for execution info

integration_execution_details: Option<IntegrationExecutionDetails>

The details about this integration execution.

execution_type: i32

Specifies whether this execution info corresponds to actual integration or test case.

execution_method: i32

The ways user posts this event.

integration_snapshot_number: i64

An increasing sequence that is set when a new snapshot (Integration Version) is created.

Implementations

impl ExecutionInfo

pub fn integration_version_state(&self) -> IntegrationState

Returns the enum value of integration_version_state, or the default if the field is set to an invalid enum value.

pub fn set_integration_version_state(&mut self, value: IntegrationState)

Sets integration_version_state to the provided enum value.

pub fn execution_type(&self) -> ExecutionType

Returns the enum value of execution_type, or the default if the field is set to an invalid enum value.

pub fn set_execution_type(&mut self, value: ExecutionType)

Sets execution_type to the provided enum value.

pub fn execution_method(&self) -> ExecutionMethod

Returns the enum value of execution_method, or the default if the field is set to an invalid enum value.

pub fn set_execution_method(&mut self, value: ExecutionMethod)

Sets execution_method to the provided enum value.