Struct google_api_proto::google::cloud::iot::v1::device_manager_client::DeviceManagerClient

pub struct DeviceManagerClient<T> { /* private fields */ }

Internet of Things (IoT) service. Securely connect and manage IoT devices.

Implementations

impl<T> DeviceManagerClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> DeviceManagerClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_device_registry( &mut self, request: impl IntoRequest<CreateDeviceRegistryRequest>, ) -> Result<Response<DeviceRegistry>, Status>

Creates a device registry that contains devices.

pub async fn get_device_registry( &mut self, request: impl IntoRequest<GetDeviceRegistryRequest>, ) -> Result<Response<DeviceRegistry>, Status>

Gets a device registry configuration.

pub async fn update_device_registry( &mut self, request: impl IntoRequest<UpdateDeviceRegistryRequest>, ) -> Result<Response<DeviceRegistry>, Status>

Updates a device registry configuration.

pub async fn delete_device_registry( &mut self, request: impl IntoRequest<DeleteDeviceRegistryRequest>, ) -> Result<Response<()>, Status>

Deletes a device registry configuration.

pub async fn list_device_registries( &mut self, request: impl IntoRequest<ListDeviceRegistriesRequest>, ) -> Result<Response<ListDeviceRegistriesResponse>, Status>

Lists device registries.

pub async fn create_device( &mut self, request: impl IntoRequest<CreateDeviceRequest>, ) -> Result<Response<Device>, Status>

Creates a device in a device registry.

pub async fn get_device( &mut self, request: impl IntoRequest<GetDeviceRequest>, ) -> Result<Response<Device>, Status>

Gets details about a device.

pub async fn update_device( &mut self, request: impl IntoRequest<UpdateDeviceRequest>, ) -> Result<Response<Device>, Status>

Updates a device.

pub async fn delete_device( &mut self, request: impl IntoRequest<DeleteDeviceRequest>, ) -> Result<Response<()>, Status>

Deletes a device.

pub async fn list_devices( &mut self, request: impl IntoRequest<ListDevicesRequest>, ) -> Result<Response<ListDevicesResponse>, Status>

List devices in a device registry.

pub async fn modify_cloud_to_device_config( &mut self, request: impl IntoRequest<ModifyCloudToDeviceConfigRequest>, ) -> Result<Response<DeviceConfig>, Status>

Modifies the configuration for the device, which is eventually sent from the Cloud IoT Core servers. Returns the modified configuration version and its metadata.

pub async fn list_device_config_versions( &mut self, request: impl IntoRequest<ListDeviceConfigVersionsRequest>, ) -> Result<Response<ListDeviceConfigVersionsResponse>, Status>

Lists the last few versions of the device configuration in descending order (i.e.: newest first).

pub async fn list_device_states( &mut self, request: impl IntoRequest<ListDeviceStatesRequest>, ) -> Result<Response<ListDeviceStatesResponse>, Status>

Lists the last few versions of the device state in descending order (i.e.: newest first).

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Sets the access control policy on the specified resource. Replaces any existing policy.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest>, ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error.

pub async fn send_command_to_device( &mut self, request: impl IntoRequest<SendCommandToDeviceRequest>, ) -> Result<Response<SendCommandToDeviceResponse>, Status>

Sends a command to the specified device. In order for a device to be able to receive commands, it must:

be connected to Cloud IoT Core using the MQTT protocol, and

be subscribed to the group of MQTT topics specified by /devices/{device-id}/commands/#. This subscription will receive commands at the top-level topic /devices/{device-id}/commands as well as commands for subfolders, like /devices/{device-id}/commands/subfolder. Note that subscribing to specific subfolders is not supported. If the command could not be delivered to the device, this method will return an error; in particular, if the device is not subscribed, this method will return FAILED_PRECONDITION. Otherwise, this method will return OK. If the subscription is QoS 1, at least once delivery will be guaranteed; for QoS 0, no acknowledgment will be expected from the device.

pub async fn bind_device_to_gateway( &mut self, request: impl IntoRequest<BindDeviceToGatewayRequest>, ) -> Result<Response<BindDeviceToGatewayResponse>, Status>

Associates the device with the gateway.

pub async fn unbind_device_from_gateway( &mut self, request: impl IntoRequest<UnbindDeviceFromGatewayRequest>, ) -> Result<Response<UnbindDeviceFromGatewayResponse>, Status>

Deletes the association between the device and the gateway.