Enum google_api_proto::google::cloud::language::v1beta1::dependency_edge::Label

#[repr(i32)]
pub enum Label {

    Unknown = 0,
    Abbrev = 1,
    Acomp = 2,
    Advcl = 3,
    Advmod = 4,
    Amod = 5,
    Appos = 6,
    Attr = 7,
    Aux = 8,
    Auxpass = 9,
    Cc = 10,
    Ccomp = 11,
    Conj = 12,
    Csubj = 13,
    Csubjpass = 14,
    Dep = 15,
    Det = 16,
    Discourse = 17,
    Dobj = 18,
    Expl = 19,
    Goeswith = 20,
    Iobj = 21,
    Mark = 22,
    Mwe = 23,
    Mwv = 24,
    Neg = 25,
    Nn = 26,
    Npadvmod = 27,
    Nsubj = 28,
    Nsubjpass = 29,
    Num = 30,
    Number = 31,
    P = 32,
    Parataxis = 33,
    Partmod = 34,
    Pcomp = 35,
    Pobj = 36,
    Poss = 37,
    Postneg = 38,
    Precomp = 39,
    Preconj = 40,
    Predet = 41,
    Pref = 42,
    Prep = 43,
    Pronl = 44,
    Prt = 45,
    Ps = 46,
    Quantmod = 47,
    Rcmod = 48,
    Rcmodrel = 49,
    Rdrop = 50,
    Ref = 51,
    Remnant = 52,
    Reparandum = 53,
    Root = 54,
    Snum = 55,
    Suff = 56,
    Tmod = 57,
    Topic = 58,
    Vmod = 59,
    Vocative = 60,
    Xcomp = 61,
    Suffix = 62,
    Title = 63,
    Advphmod = 64,
    Auxcaus = 65,
    Auxvv = 66,
    Dtmod = 67,
    Foreign = 68,
    Kw = 69,
    List = 70,
    Nomc = 71,
    Nomcsubj = 72,
    Nomcsubjpass = 73,
    Numc = 74,
    Cop = 75,
    Dislocated = 76,

}

The parse label enum for the token.

Variants

Unknown = 0

Unknown

Abbrev = 1

Abbreviation modifier

Acomp = 2

Adjectival complement

Advcl = 3

Adverbial clause modifier

Advmod = 4

Adverbial modifier

Amod = 5

Adjectival modifier of an NP

Appos = 6

Appositional modifier of an NP

Attr = 7

Attribute dependent of a copular verb

Aux = 8

Auxiliary (non-main) verb

Auxpass = 9

Passive auxiliary

Cc = 10

Coordinating conjunction

Ccomp = 11

Clausal complement of a verb or adjective

Conj = 12

Conjunct

Csubj = 13

Clausal subject

Csubjpass = 14

Clausal passive subject

Dep = 15

Dependency (unable to determine)

Det = 16

Determiner

Discourse = 17

Discourse

Dobj = 18

Direct object

Expl = 19

Expletive

Goeswith = 20

Goes with (part of a word in a text not well edited)

Iobj = 21

Indirect object

Mark = 22

Marker (word introducing a subordinate clause)

Mwe = 23

Multi-word expression

Mwv = 24

Multi-word verbal expression

Neg = 25

Negation modifier

Nn = 26

Noun compound modifier

Npadvmod = 27

Noun phrase used as an adverbial modifier

Nsubj = 28

Nominal subject

Nsubjpass = 29

Passive nominal subject

Num = 30

Numeric modifier of a noun

Number = 31

Element of compound number

P = 32

Punctuation mark

Parataxis = 33

Parataxis relation

Partmod = 34

Participial modifier

Pcomp = 35

The complement of a preposition is a clause

Pobj = 36

Object of a preposition

Poss = 37

Possession modifier

Postneg = 38

Postverbal negative particle

Precomp = 39

Predicate complement

Preconj = 40

Preconjunt

Predet = 41

Predeterminer

Pref = 42

Prefix

Prep = 43

Prepositional modifier

Pronl = 44

The relationship between a verb and verbal morpheme

Prt = 45

Particle

Ps = 46

Associative or possessive marker

Quantmod = 47

Quantifier phrase modifier

Rcmod = 48

Relative clause modifier

Rcmodrel = 49

Complementizer in relative clause

Rdrop = 50

Ellipsis without a preceding predicate

Ref = 51

Referent

Remnant = 52

Remnant

Reparandum = 53

Reparandum

Root = 54

Root

Snum = 55

Suffix specifying a unit of number

Suff = 56

Suffix

Tmod = 57

Temporal modifier

Topic = 58

Topic marker

Vmod = 59

Clause headed by an infinite form of the verb that modifies a noun

Vocative = 60

Vocative

Xcomp = 61

Open clausal complement

Suffix = 62

Name suffix

Title = 63

Name title

Advphmod = 64

Adverbial phrase modifier

Auxcaus = 65

Causative auxiliary

Auxvv = 66

Helper auxiliary

Dtmod = 67

Rentaishi (Prenominal modifier)

Foreign = 68

Foreign words

Kw = 69

Keyword

List = 70

List for chains of comparable items

Nomc = 71

Nominalized clause

Nomcsubj = 72

Nominalized clausal subject

Nomcsubjpass = 73

Nominalized clausal passive

Numc = 74

Compound of numeric modifier

Cop = 75

Copula

Dislocated = 76

Dislocated relation (for fronted/topicalized elements)

Implementations

impl Label

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Label.

pub fn from_i32(value: i32) -> Option<Label>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Label, or None if value is not a valid variant.

impl Label

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.