Module google_api_proto::google::cloud::language::v2

Modules

annotate_text_request

Nested message and enum types in AnnotateTextRequest.

document

Nested message and enum types in Document.

entity

Nested message and enum types in Entity.

entity_mention

Nested message and enum types in EntityMention.

language_service_client

Generated client implementations.

moderate_text_request

Nested message and enum types in ModerateTextRequest.

Structs

AnalyzeEntitiesRequest

The entity analysis request message.

AnalyzeEntitiesResponse

The entity analysis response message.

AnalyzeSentimentRequest

The sentiment analysis request message.

AnalyzeSentimentResponse

The sentiment analysis response message.

AnnotateTextRequest

The request message for the text annotation API, which can perform multiple analysis types in one call.

AnnotateTextResponse

The text annotations response message.

ClassificationCategory

Represents a category returned from the text classifier.

ClassifyTextRequest

The document classification request message.

ClassifyTextResponse

The document classification response message.

Document

Represents the input to API methods.

Entity

Represents a phrase in the text that is a known entity, such as a person, an organization, or location. The API associates information, such as probability and mentions, with entities.

EntityMention

Represents a mention for an entity in the text. Currently, proper noun mentions are supported.

ModerateTextRequest

The document moderation request message.

ModerateTextResponse

The document moderation response message.

Sentence

Represents a sentence in the input document.

Sentiment

Represents the feeling associated with the entire text or entities in the text.

TextSpan

Represents a text span in the input document.

Enums

EncodingType

Represents the text encoding that the caller uses to process the output. Providing an EncodingType is recommended because the API provides the beginning offsets for various outputs, such as tokens and mentions, and languages that natively use different text encodings may access offsets differently.