Module google_api_proto::google::cloud::managedidentities::v1beta1

Modules

domain

Nested message and enum types in Domain.

managed_identities_service_client

Generated client implementations.

trust

Nested message and enum types in Trust.

Structs

AttachTrustRequest

Request message for [AttachTrust][google.cloud.managedidentities.v1beta1.AttachTrust]

CreateMicrosoftAdDomainRequest

Request message for [CreateMicrosoftAdDomain][google.cloud.managedidentities.v1beta1.CreateMicrosoftAdDomain]

DeleteDomainRequest

Request message for [DeleteDomain][google.cloud.managedidentities.v1beta1.DeleteDomain]

DetachTrustRequest

Request message for [DetachTrust][google.cloud.managedidentities.v1beta1.DetachTrust]

Domain

Represents a managed Microsoft Active Directory domain. If the domain is being changed, it will be placed into the UPDATING state, which indicates that the resource is being reconciled. At this point, Get will reflect an intermediate state.

GetDomainRequest

Request message for [GetDomain][google.cloud.managedidentities.v1beta1.GetDomain]

ListDomainsRequest

Request message for [ListDomains][google.cloud.managedidentities.v1beta1.ListDomains]

ListDomainsResponse

Response message for [ListDomains][google.cloud.managedidentities.v1beta1.ListDomains]

OpMetadata

Represents the metadata of the long-running operation.

ReconfigureTrustRequest

Request message for [ReconfigureTrust][google.cloud.managedidentities.v1beta1.ReconfigureTrust]

ResetAdminPasswordRequest

Request message for [ResetAdminPassword][google.cloud.managedidentities.v1beta1.ResetAdminPassword]

ResetAdminPasswordResponse

Response message for [ResetAdminPassword][google.cloud.managedidentities.v1beta1.ResetAdminPassword]

Trust

Represents a relationship between two domains. This allows a controller in one domain to authenticate a user in another domain.

UpdateDomainRequest

Request message for [UpdateDomain][google.cloud.managedidentities.v1beta1.UpdateDomain]

ValidateTrustRequest

Request message for [ValidateTrust][google.cloud.managedidentities.v1beta1.ValidateTrust]