Module google_api_proto::google::cloud::metastore::v1alpha

Modules

backend_metastore

Nested message and enum types in BackendMetastore.

backup

Nested message and enum types in Backup.

database_dump_spec

Nested message and enum types in DatabaseDumpSpec.

dataproc_metastore_client

Generated client implementations.

dataproc_metastore_federation_client

Generated client implementations.

export_metadata_request

Nested message and enum types in ExportMetadataRequest.

federation

Nested message and enum types in Federation.

hive_metastore_config

Nested message and enum types in HiveMetastoreConfig.

location_metadata

Nested message and enum types in LocationMetadata.

metadata_export

Nested message and enum types in MetadataExport.

metadata_import

Nested message and enum types in MetadataImport.

network_config

Nested message and enum types in NetworkConfig.

restore

Nested message and enum types in Restore.

scaling_config

Nested message and enum types in ScalingConfig.

secret

Nested message and enum types in Secret.

service

Nested message and enum types in Service.

telemetry_config

Nested message and enum types in TelemetryConfig.

Structs

AlterMetadataResourceLocationRequest

Request message for [DataprocMetastore.AlterMetadataResourceLocation][google.cloud.metastore.v1alpha.DataprocMetastore.AlterMetadataResourceLocation].

AlterMetadataResourceLocationResponse

Response message for [DataprocMetastore.AlterMetadataResourceLocation][google.cloud.metastore.v1alpha.DataprocMetastore.AlterMetadataResourceLocation].

AuxiliaryVersionConfig

Configuration information for the auxiliary service versions.

BackendMetastore

Represents a backend metastore for the federation.

Backup

The details of a backup resource.

CreateBackupRequest

Request message for [DataprocMetastore.CreateBackup][google.cloud.metastore.v1alpha.DataprocMetastore.CreateBackup].

CreateFederationRequest

Request message for CreateFederation.

CreateMetadataImportRequest

Request message for [DataprocMetastore.CreateMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.CreateMetadataImport].

CreateServiceRequest

Request message for [DataprocMetastore.CreateService][google.cloud.metastore.v1alpha.DataprocMetastore.CreateService].

DataCatalogConfig

Specifies how metastore metadata should be integrated with the Data Catalog service.

DatabaseDumpSpec

The specification of database dump to import from or export to.

DataplexConfig

Specifies how metastore metadata should be integrated with the Dataplex service.

DeleteBackupRequest

Request message for [DataprocMetastore.DeleteBackup][google.cloud.metastore.v1alpha.DataprocMetastore.DeleteBackup].

DeleteFederationRequest

Request message for DeleteFederation.

DeleteServiceRequest

Request message for [DataprocMetastore.DeleteService][google.cloud.metastore.v1alpha.DataprocMetastore.DeleteService].

EncryptionConfig

Encryption settings for the service.

ErrorDetails

Error details in public error message for [DataprocMetastore.QueryMetadata][google.cloud.metastore.v1alpha.DataprocMetastore.QueryMetadata].

ExportMetadataRequest

Request message for [DataprocMetastore.ExportMetadata][google.cloud.metastore.v1alpha.DataprocMetastore.ExportMetadata].

Federation

Represents a federation of multiple backend metastores.

GetBackupRequest

Request message for [DataprocMetastore.GetBackup][google.cloud.metastore.v1alpha.DataprocMetastore.GetBackup].

GetFederationRequest

Request message for GetFederation.

GetMetadataImportRequest

Request message for [DataprocMetastore.GetMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.GetMetadataImport].

GetServiceRequest

Request message for [DataprocMetastore.GetService][google.cloud.metastore.v1alpha.DataprocMetastore.GetService].

HiveMetastoreConfig

Specifies configuration information specific to running Hive metastore software as the metastore service.

KerberosConfig

Configuration information for a Kerberos principal.

Lake

Represents a Lake resource

ListBackupsRequest

Request message for [DataprocMetastore.ListBackups][google.cloud.metastore.v1alpha.DataprocMetastore.ListBackups].

ListBackupsResponse

Response message for [DataprocMetastore.ListBackups][google.cloud.metastore.v1alpha.DataprocMetastore.ListBackups].

ListFederationsRequest

Request message for ListFederations.

ListFederationsResponse

Response message for ListFederations

ListMetadataImportsRequest

Request message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1alpha.DataprocMetastore.ListMetadataImports].

ListMetadataImportsResponse

Response message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1alpha.DataprocMetastore.ListMetadataImports].

ListServicesRequest

Request message for [DataprocMetastore.ListServices][google.cloud.metastore.v1alpha.DataprocMetastore.ListServices].

ListServicesResponse

Response message for [DataprocMetastore.ListServices][google.cloud.metastore.v1alpha.DataprocMetastore.ListServices].

LocationMetadata

Metadata about the service in a location.

MaintenanceWindow

Maintenance window. This specifies when Dataproc Metastore may perform system maintenance operation to the service.

MetadataExport

The details of a metadata export operation.

MetadataImport

A metastore resource that imports metadata.

MetadataIntegration

Specifies how metastore metadata should be integrated with external services.

MetadataManagementActivity

The metadata management activities of the metastore service.

MoveTableToDatabaseRequest

Request message for [DataprocMetastore.MoveTableToDatabase][google.cloud.metastore.v1alpha.DataprocMetastore.MoveTableToDatabase].

MoveTableToDatabaseResponse

Response message for [DataprocMetastore.MoveTableToDatabase][google.cloud.metastore.v1alpha.DataprocMetastore.MoveTableToDatabase].

NetworkConfig

Network configuration for the Dataproc Metastore service.

OperationMetadata

Represents the metadata of a long-running operation.

QueryMetadataRequest

Request message for [DataprocMetastore.QueryMetadata][google.cloud.metastore.v1alpha.DataprocMetastore.QueryMetadata].

QueryMetadataResponse

Response message for [DataprocMetastore.QueryMetadata][google.cloud.metastore.v1alpha.DataprocMetastore.QueryMetadata].

RemoveIamPolicyRequest

Request message for [DataprocMetastore.RemoveIamPolicy][google.cloud.metastore.v1alpha.DataprocMetastore.RemoveIamPolicy].

RemoveIamPolicyResponse

Response message for [DataprocMetastore.RemoveIamPolicy][google.cloud.metastore.v1alpha.DataprocMetastore.RemoveIamPolicy].

Restore

The details of a metadata restore operation.

RestoreServiceRequest

Request message for [DataprocMetastore.Restore][].

ScalingConfig

Represents the scaling configuration of a metastore service.

Secret

A securely stored value.

Service

A managed metastore service that serves metadata queries.

TelemetryConfig

Telemetry Configuration for the Dataproc Metastore service.

UpdateFederationRequest

Request message for UpdateFederation.

UpdateMetadataImportRequest

Request message for [DataprocMetastore.UpdateMetadataImport][google.cloud.metastore.v1alpha.DataprocMetastore.UpdateMetadataImport].

UpdateServiceRequest

Request message for [DataprocMetastore.UpdateService][google.cloud.metastore.v1alpha.DataprocMetastore.UpdateService].