Struct google_api_proto::google::cloud::metastore::v1alpha::dataproc_metastore_client::DataprocMetastoreClient

pub struct DataprocMetastoreClient<T> { /* private fields */ }

Configures and manages metastore services. Metastore services are fully managed, highly available, autoscaled, autohealing, OSS-native deployments of technical metadata management software. Each metastore service exposes a network endpoint through which metadata queries are served. Metadata queries can originate from a variety of sources, including Apache Hive, Apache Presto, and Apache Spark.

The Dataproc Metastore API defines the following resource model:

The service works with a collection of Google Cloud projects, named: /projects/*

Each project has a collection of available locations, named: /locations/* (a location must refer to a Google Cloud region)

Each location has a collection of services, named: /services/*

Dataproc Metastore services are resources with names of the form:

/projects/{project_number}/locations/{location_id}/services/{service_id}.

Implementations

impl<T> DataprocMetastoreClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> DataprocMetastoreClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_services( &mut self, request: impl IntoRequest<ListServicesRequest> ) -> Result<Response<ListServicesResponse>, Status>

Lists services in a project and location.

pub async fn get_service( &mut self, request: impl IntoRequest<GetServiceRequest> ) -> Result<Response<Service>, Status>

Gets the details of a single service.

pub async fn create_service( &mut self, request: impl IntoRequest<CreateServiceRequest> ) -> Result<Response<Operation>, Status>

Creates a metastore service in a project and location.

pub async fn update_service( &mut self, request: impl IntoRequest<UpdateServiceRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single service.

pub async fn delete_service( &mut self, request: impl IntoRequest<DeleteServiceRequest> ) -> Result<Response<Operation>, Status>

Deletes a single service.

pub async fn list_metadata_imports( &mut self, request: impl IntoRequest<ListMetadataImportsRequest> ) -> Result<Response<ListMetadataImportsResponse>, Status>

Lists imports in a service.

pub async fn get_metadata_import( &mut self, request: impl IntoRequest<GetMetadataImportRequest> ) -> Result<Response<MetadataImport>, Status>

Gets details of a single import.

pub async fn create_metadata_import( &mut self, request: impl IntoRequest<CreateMetadataImportRequest> ) -> Result<Response<Operation>, Status>

Creates a new MetadataImport in a given project and location.

pub async fn update_metadata_import( &mut self, request: impl IntoRequest<UpdateMetadataImportRequest> ) -> Result<Response<Operation>, Status>

Updates a single import. Only the description field of MetadataImport is supported to be updated.

pub async fn export_metadata( &mut self, request: impl IntoRequest<ExportMetadataRequest> ) -> Result<Response<Operation>, Status>

Exports metadata from a service.

pub async fn restore_service( &mut self, request: impl IntoRequest<RestoreServiceRequest> ) -> Result<Response<Operation>, Status>

Restores a service from a backup.

pub async fn list_backups( &mut self, request: impl IntoRequest<ListBackupsRequest> ) -> Result<Response<ListBackupsResponse>, Status>

Lists backups in a service.

pub async fn get_backup( &mut self, request: impl IntoRequest<GetBackupRequest> ) -> Result<Response<Backup>, Status>

Gets details of a single backup.

pub async fn create_backup( &mut self, request: impl IntoRequest<CreateBackupRequest> ) -> Result<Response<Operation>, Status>

Creates a new backup in a given project and location.

pub async fn delete_backup( &mut self, request: impl IntoRequest<DeleteBackupRequest> ) -> Result<Response<Operation>, Status>

Deletes a single backup.

pub async fn remove_iam_policy( &mut self, request: impl IntoRequest<RemoveIamPolicyRequest> ) -> Result<Response<RemoveIamPolicyResponse>, Status>

Removes the attached IAM policies for a resource

pub async fn query_metadata( &mut self, request: impl IntoRequest<QueryMetadataRequest> ) -> Result<Response<Operation>, Status>

Query DPMS metadata.

pub async fn move_table_to_database( &mut self, request: impl IntoRequest<MoveTableToDatabaseRequest> ) -> Result<Response<Operation>, Status>

Move a table to another database.

pub async fn alter_metadata_resource_location( &mut self, request: impl IntoRequest<AlterMetadataResourceLocationRequest> ) -> Result<Response<Operation>, Status>

Alter metadata resource location. The metadata resource can be a database, table, or partition. This functionality only updates the parent directory for the respective metadata resource and does not transfer any existing data to the new location.