Struct google_api_proto::google::cloud::migrationcenter::v1::migration_center_client::MigrationCenterClient

pub struct MigrationCenterClient<T> { /* private fields */ }

Service describing handlers for resources.

Implementations

impl<T> MigrationCenterClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> MigrationCenterClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_assets( &mut self, request: impl IntoRequest<ListAssetsRequest> ) -> Result<Response<ListAssetsResponse>, Status>

Lists all the assets in a given project and location.

pub async fn get_asset( &mut self, request: impl IntoRequest<GetAssetRequest> ) -> Result<Response<Asset>, Status>

Gets the details of an asset.

pub async fn update_asset( &mut self, request: impl IntoRequest<UpdateAssetRequest> ) -> Result<Response<Asset>, Status>

Updates the parameters of an asset.

pub async fn batch_update_assets( &mut self, request: impl IntoRequest<BatchUpdateAssetsRequest> ) -> Result<Response<BatchUpdateAssetsResponse>, Status>

Updates the parameters of a list of assets.

pub async fn delete_asset( &mut self, request: impl IntoRequest<DeleteAssetRequest> ) -> Result<Response<()>, Status>

Deletes an asset.

pub async fn batch_delete_assets( &mut self, request: impl IntoRequest<BatchDeleteAssetsRequest> ) -> Result<Response<()>, Status>

Deletes list of Assets.

pub async fn report_asset_frames( &mut self, request: impl IntoRequest<ReportAssetFramesRequest> ) -> Result<Response<ReportAssetFramesResponse>, Status>

Reports a set of frames.

pub async fn aggregate_assets_values( &mut self, request: impl IntoRequest<AggregateAssetsValuesRequest> ) -> Result<Response<AggregateAssetsValuesResponse>, Status>

Aggregates the requested fields based on provided function.

pub async fn create_import_job( &mut self, request: impl IntoRequest<CreateImportJobRequest> ) -> Result<Response<Operation>, Status>

Creates an import job.

pub async fn list_import_jobs( &mut self, request: impl IntoRequest<ListImportJobsRequest> ) -> Result<Response<ListImportJobsResponse>, Status>

Lists all import jobs.

pub async fn get_import_job( &mut self, request: impl IntoRequest<GetImportJobRequest> ) -> Result<Response<ImportJob>, Status>

Gets the details of an import job.

pub async fn delete_import_job( &mut self, request: impl IntoRequest<DeleteImportJobRequest> ) -> Result<Response<Operation>, Status>

Deletes an import job.

pub async fn update_import_job( &mut self, request: impl IntoRequest<UpdateImportJobRequest> ) -> Result<Response<Operation>, Status>

Updates an import job.

pub async fn validate_import_job( &mut self, request: impl IntoRequest<ValidateImportJobRequest> ) -> Result<Response<Operation>, Status>

Validates an import job.

pub async fn run_import_job( &mut self, request: impl IntoRequest<RunImportJobRequest> ) -> Result<Response<Operation>, Status>

Runs an import job.

pub async fn get_import_data_file( &mut self, request: impl IntoRequest<GetImportDataFileRequest> ) -> Result<Response<ImportDataFile>, Status>

Gets an import data file.

pub async fn list_import_data_files( &mut self, request: impl IntoRequest<ListImportDataFilesRequest> ) -> Result<Response<ListImportDataFilesResponse>, Status>

List import data files.

pub async fn create_import_data_file( &mut self, request: impl IntoRequest<CreateImportDataFileRequest> ) -> Result<Response<Operation>, Status>

Creates an import data file.

pub async fn delete_import_data_file( &mut self, request: impl IntoRequest<DeleteImportDataFileRequest> ) -> Result<Response<Operation>, Status>

Delete an import data file.

pub async fn list_groups( &mut self, request: impl IntoRequest<ListGroupsRequest> ) -> Result<Response<ListGroupsResponse>, Status>

Lists all groups in a given project and location.

pub async fn get_group( &mut self, request: impl IntoRequest<GetGroupRequest> ) -> Result<Response<Group>, Status>

Gets the details of a group.

pub async fn create_group( &mut self, request: impl IntoRequest<CreateGroupRequest> ) -> Result<Response<Operation>, Status>

Creates a new group in a given project and location.

pub async fn update_group( &mut self, request: impl IntoRequest<UpdateGroupRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a group.

pub async fn delete_group( &mut self, request: impl IntoRequest<DeleteGroupRequest> ) -> Result<Response<Operation>, Status>

Deletes a group.

pub async fn add_assets_to_group( &mut self, request: impl IntoRequest<AddAssetsToGroupRequest> ) -> Result<Response<Operation>, Status>

Adds assets to a group.

pub async fn remove_assets_from_group( &mut self, request: impl IntoRequest<RemoveAssetsFromGroupRequest> ) -> Result<Response<Operation>, Status>

Removes assets from a group.

pub async fn list_error_frames( &mut self, request: impl IntoRequest<ListErrorFramesRequest> ) -> Result<Response<ListErrorFramesResponse>, Status>

Lists all error frames in a given source and location.

pub async fn get_error_frame( &mut self, request: impl IntoRequest<GetErrorFrameRequest> ) -> Result<Response<ErrorFrame>, Status>

Gets the details of an error frame.

pub async fn list_sources( &mut self, request: impl IntoRequest<ListSourcesRequest> ) -> Result<Response<ListSourcesResponse>, Status>

Lists all the sources in a given project and location.

pub async fn get_source( &mut self, request: impl IntoRequest<GetSourceRequest> ) -> Result<Response<Source>, Status>

Gets the details of a source.

pub async fn create_source( &mut self, request: impl IntoRequest<CreateSourceRequest> ) -> Result<Response<Operation>, Status>

Creates a new source in a given project and location.

pub async fn update_source( &mut self, request: impl IntoRequest<UpdateSourceRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a source.

pub async fn delete_source( &mut self, request: impl IntoRequest<DeleteSourceRequest> ) -> Result<Response<Operation>, Status>

Deletes a source.

pub async fn list_preference_sets( &mut self, request: impl IntoRequest<ListPreferenceSetsRequest> ) -> Result<Response<ListPreferenceSetsResponse>, Status>

Lists all the preference sets in a given project and location.

pub async fn get_preference_set( &mut self, request: impl IntoRequest<GetPreferenceSetRequest> ) -> Result<Response<PreferenceSet>, Status>

Gets the details of a preference set.

pub async fn create_preference_set( &mut self, request: impl IntoRequest<CreatePreferenceSetRequest> ) -> Result<Response<Operation>, Status>

Creates a new preference set in a given project and location.

pub async fn update_preference_set( &mut self, request: impl IntoRequest<UpdatePreferenceSetRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a preference set.

pub async fn delete_preference_set( &mut self, request: impl IntoRequest<DeletePreferenceSetRequest> ) -> Result<Response<Operation>, Status>

Deletes a preference set.

pub async fn get_settings( &mut self, request: impl IntoRequest<GetSettingsRequest> ) -> Result<Response<Settings>, Status>

Gets the details of regional settings.

pub async fn update_settings( &mut self, request: impl IntoRequest<UpdateSettingsRequest> ) -> Result<Response<Operation>, Status>

Updates the regional-level project settings.

pub async fn create_report_config( &mut self, request: impl IntoRequest<CreateReportConfigRequest> ) -> Result<Response<Operation>, Status>

Creates a report configuration.

pub async fn get_report_config( &mut self, request: impl IntoRequest<GetReportConfigRequest> ) -> Result<Response<ReportConfig>, Status>

Gets details of a single ReportConfig.

pub async fn list_report_configs( &mut self, request: impl IntoRequest<ListReportConfigsRequest> ) -> Result<Response<ListReportConfigsResponse>, Status>

Lists ReportConfigs in a given project and location.

pub async fn delete_report_config( &mut self, request: impl IntoRequest<DeleteReportConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes a ReportConfig.

pub async fn create_report( &mut self, request: impl IntoRequest<CreateReportRequest> ) -> Result<Response<Operation>, Status>

Creates a report.

pub async fn get_report( &mut self, request: impl IntoRequest<GetReportRequest> ) -> Result<Response<Report>, Status>

Gets details of a single Report.

pub async fn list_reports( &mut self, request: impl IntoRequest<ListReportsRequest> ) -> Result<Response<ListReportsResponse>, Status>

Lists Reports in a given ReportConfig.

pub async fn delete_report( &mut self, request: impl IntoRequest<DeleteReportRequest> ) -> Result<Response<Operation>, Status>

Deletes a Report.