Struct google_api_proto::google::cloud::migrationcenter::v1::VirtualMachinePreferences

pub struct VirtualMachinePreferences {
    pub target_product: i32,
    pub region_preferences: Option<RegionPreferences>,
    pub commitment_plan: i32,
    pub sizing_optimization_strategy: i32,
    pub compute_engine_preferences: Option<ComputeEnginePreferences>,
    pub vmware_engine_preferences: Option<VmwareEnginePreferences>,
    pub sole_tenancy_preferences: Option<SoleTenancyPreferences>,
}

VirtualMachinePreferences enables you to create sets of assumptions, for example, a geographical location and pricing track, for your migrated virtual machines. The set of preferences influence recommendations for migrating virtual machine assets.

Fields
target_product: i32

Target product for assets using this preference set. Specify either target product or business goal, but not both.

region_preferences: Option<RegionPreferences>

Region preferences for assets using this preference set. If you are unsure which value to set, the migration service API region is often a good value to start with.

commitment_plan: i32

Commitment plan to consider when calculating costs for virtual machine insights and recommendations. If you are unsure which value to set, a 3 year commitment plan is often a good value to start with.

sizing_optimization_strategy: i32

Sizing optimization strategy specifies the preferred strategy used when extrapolating usage data to calculate insights and recommendations for a virtual machine. If you are unsure which value to set, a moderate sizing optimization strategy is often a good value to start with.

compute_engine_preferences: Option<ComputeEnginePreferences>

Compute Engine preferences concern insights and recommendations for Compute Engine target.

vmware_engine_preferences: Option<VmwareEnginePreferences>

Preferences concerning insights and recommendations for Google Cloud VMware Engine.

sole_tenancy_preferences: Option<SoleTenancyPreferences>

Preferences concerning Sole Tenant nodes and virtual machines.

Implementations

impl VirtualMachinePreferences

pub fn target_product(&self) -> ComputeMigrationTargetProduct

Returns the enum value of target_product, or the default if the field is set to an invalid enum value.

pub fn set_target_product(&mut self, value: ComputeMigrationTargetProduct)

Sets target_product to the provided enum value.

pub fn commitment_plan(&self) -> CommitmentPlan

Returns the enum value of commitment_plan, or the default if the field is set to an invalid enum value.

pub fn set_commitment_plan(&mut self, value: CommitmentPlan)

Sets commitment_plan to the provided enum value.

pub fn sizing_optimization_strategy(&self) -> SizingOptimizationStrategy

Returns the enum value of sizing_optimization_strategy, or the default if the field is set to an invalid enum value.

pub fn set_sizing_optimization_strategy( &mut self, value: SizingOptimizationStrategy )

Sets sizing_optimization_strategy to the provided enum value.