Struct google_api_proto::google::cloud::netapp::v1::net_app_client::NetAppClient

pub struct NetAppClient<T> { /* private fields */ }

NetApp Files Google Cloud Service

Implementations

impl<T> NetAppClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> NetAppClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_storage_pools( &mut self, request: impl IntoRequest<ListStoragePoolsRequest>, ) -> Result<Response<ListStoragePoolsResponse>, Status>

Returns descriptions of all storage pools owned by the caller.

pub async fn create_storage_pool( &mut self, request: impl IntoRequest<CreateStoragePoolRequest>, ) -> Result<Response<Operation>, Status>

Creates a new storage pool.

pub async fn get_storage_pool( &mut self, request: impl IntoRequest<GetStoragePoolRequest>, ) -> Result<Response<StoragePool>, Status>

Returns the description of the specified storage pool by poolId.

pub async fn update_storage_pool( &mut self, request: impl IntoRequest<UpdateStoragePoolRequest>, ) -> Result<Response<Operation>, Status>

Updates the storage pool properties with the full spec

pub async fn delete_storage_pool( &mut self, request: impl IntoRequest<DeleteStoragePoolRequest>, ) -> Result<Response<Operation>, Status>

Warning! This operation will permanently delete the storage pool.

pub async fn list_volumes( &mut self, request: impl IntoRequest<ListVolumesRequest>, ) -> Result<Response<ListVolumesResponse>, Status>

Lists Volumes in a given project.

pub async fn get_volume( &mut self, request: impl IntoRequest<GetVolumeRequest>, ) -> Result<Response<Volume>, Status>

Gets details of a single Volume.

pub async fn create_volume( &mut self, request: impl IntoRequest<CreateVolumeRequest>, ) -> Result<Response<Operation>, Status>

Creates a new Volume in a given project and location.

pub async fn update_volume( &mut self, request: impl IntoRequest<UpdateVolumeRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Volume.

pub async fn delete_volume( &mut self, request: impl IntoRequest<DeleteVolumeRequest>, ) -> Result<Response<Operation>, Status>

Deletes a single Volume.

pub async fn revert_volume( &mut self, request: impl IntoRequest<RevertVolumeRequest>, ) -> Result<Response<Operation>, Status>

Revert an existing volume to a specified snapshot. Warning! This operation will permanently revert all changes made after the snapshot was created.

pub async fn list_snapshots( &mut self, request: impl IntoRequest<ListSnapshotsRequest>, ) -> Result<Response<ListSnapshotsResponse>, Status>

Returns descriptions of all snapshots for a volume.

pub async fn get_snapshot( &mut self, request: impl IntoRequest<GetSnapshotRequest>, ) -> Result<Response<Snapshot>, Status>

Describe a snapshot for a volume.

pub async fn create_snapshot( &mut self, request: impl IntoRequest<CreateSnapshotRequest>, ) -> Result<Response<Operation>, Status>

Create a new snapshot for a volume.

pub async fn delete_snapshot( &mut self, request: impl IntoRequest<DeleteSnapshotRequest>, ) -> Result<Response<Operation>, Status>

Deletes a snapshot.

pub async fn update_snapshot( &mut self, request: impl IntoRequest<UpdateSnapshotRequest>, ) -> Result<Response<Operation>, Status>

Updates the settings of a specific snapshot.

pub async fn list_active_directories( &mut self, request: impl IntoRequest<ListActiveDirectoriesRequest>, ) -> Result<Response<ListActiveDirectoriesResponse>, Status>

Lists active directories.

pub async fn get_active_directory( &mut self, request: impl IntoRequest<GetActiveDirectoryRequest>, ) -> Result<Response<ActiveDirectory>, Status>

Describes a specified active directory.

pub async fn create_active_directory( &mut self, request: impl IntoRequest<CreateActiveDirectoryRequest>, ) -> Result<Response<Operation>, Status>

CreateActiveDirectory Creates the active directory specified in the request.

pub async fn update_active_directory( &mut self, request: impl IntoRequest<UpdateActiveDirectoryRequest>, ) -> Result<Response<Operation>, Status>

Update the parameters of an active directories.

pub async fn delete_active_directory( &mut self, request: impl IntoRequest<DeleteActiveDirectoryRequest>, ) -> Result<Response<Operation>, Status>

Delete the active directory specified in the request.

pub async fn list_kms_configs( &mut self, request: impl IntoRequest<ListKmsConfigsRequest>, ) -> Result<Response<ListKmsConfigsResponse>, Status>

Returns descriptions of all KMS configs owned by the caller.

pub async fn create_kms_config( &mut self, request: impl IntoRequest<CreateKmsConfigRequest>, ) -> Result<Response<Operation>, Status>

Creates a new KMS config.

pub async fn get_kms_config( &mut self, request: impl IntoRequest<GetKmsConfigRequest>, ) -> Result<Response<KmsConfig>, Status>

Returns the description of the specified KMS config by kms_config_id.

pub async fn update_kms_config( &mut self, request: impl IntoRequest<UpdateKmsConfigRequest>, ) -> Result<Response<Operation>, Status>

Updates the Kms config properties with the full spec

pub async fn encrypt_volumes( &mut self, request: impl IntoRequest<EncryptVolumesRequest>, ) -> Result<Response<Operation>, Status>

Encrypt the existing volumes without CMEK encryption with the desired the KMS config for the whole region.

pub async fn verify_kms_config( &mut self, request: impl IntoRequest<VerifyKmsConfigRequest>, ) -> Result<Response<VerifyKmsConfigResponse>, Status>

Verifies KMS config reachability.

pub async fn delete_kms_config( &mut self, request: impl IntoRequest<DeleteKmsConfigRequest>, ) -> Result<Response<Operation>, Status>

Warning! This operation will permanently delete the Kms config.

pub async fn list_replications( &mut self, request: impl IntoRequest<ListReplicationsRequest>, ) -> Result<Response<ListReplicationsResponse>, Status>

Returns descriptions of all replications for a volume.

pub async fn get_replication( &mut self, request: impl IntoRequest<GetReplicationRequest>, ) -> Result<Response<Replication>, Status>

Describe a replication for a volume.

pub async fn create_replication( &mut self, request: impl IntoRequest<CreateReplicationRequest>, ) -> Result<Response<Operation>, Status>

Create a new replication for a volume.

pub async fn delete_replication( &mut self, request: impl IntoRequest<DeleteReplicationRequest>, ) -> Result<Response<Operation>, Status>

Deletes a replication.

pub async fn update_replication( &mut self, request: impl IntoRequest<UpdateReplicationRequest>, ) -> Result<Response<Operation>, Status>

Updates the settings of a specific replication.

pub async fn stop_replication( &mut self, request: impl IntoRequest<StopReplicationRequest>, ) -> Result<Response<Operation>, Status>

Stop Cross Region Replication.

pub async fn resume_replication( &mut self, request: impl IntoRequest<ResumeReplicationRequest>, ) -> Result<Response<Operation>, Status>

Resume Cross Region Replication.

pub async fn reverse_replication_direction( &mut self, request: impl IntoRequest<ReverseReplicationDirectionRequest>, ) -> Result<Response<Operation>, Status>

Reverses direction of replication. Source becomes destination and destination becomes source.

pub async fn create_backup_vault( &mut self, request: impl IntoRequest<CreateBackupVaultRequest>, ) -> Result<Response<Operation>, Status>

Creates new backup vault

pub async fn get_backup_vault( &mut self, request: impl IntoRequest<GetBackupVaultRequest>, ) -> Result<Response<BackupVault>, Status>

Returns the description of the specified backup vault

pub async fn list_backup_vaults( &mut self, request: impl IntoRequest<ListBackupVaultsRequest>, ) -> Result<Response<ListBackupVaultsResponse>, Status>

Returns list of all available backup vaults.

pub async fn update_backup_vault( &mut self, request: impl IntoRequest<UpdateBackupVaultRequest>, ) -> Result<Response<Operation>, Status>

Updates the settings of a specific backup vault.

pub async fn delete_backup_vault( &mut self, request: impl IntoRequest<DeleteBackupVaultRequest>, ) -> Result<Response<Operation>, Status>

Warning! This operation will permanently delete the backup vault.

pub async fn create_backup( &mut self, request: impl IntoRequest<CreateBackupRequest>, ) -> Result<Response<Operation>, Status>

Creates a backup from the volume specified in the request The backup can be created from the given snapshot if specified in the request. If no snapshot specified, there’ll be a new snapshot taken to initiate the backup creation.

pub async fn get_backup( &mut self, request: impl IntoRequest<GetBackupRequest>, ) -> Result<Response<Backup>, Status>

Returns the description of the specified backup

pub async fn list_backups( &mut self, request: impl IntoRequest<ListBackupsRequest>, ) -> Result<Response<ListBackupsResponse>, Status>

Returns descriptions of all backups for a backupVault.

pub async fn delete_backup( &mut self, request: impl IntoRequest<DeleteBackupRequest>, ) -> Result<Response<Operation>, Status>

Warning! This operation will permanently delete the backup.

pub async fn update_backup( &mut self, request: impl IntoRequest<UpdateBackupRequest>, ) -> Result<Response<Operation>, Status>

Update backup with full spec.

pub async fn create_backup_policy( &mut self, request: impl IntoRequest<CreateBackupPolicyRequest>, ) -> Result<Response<Operation>, Status>

Creates new backup policy

pub async fn get_backup_policy( &mut self, request: impl IntoRequest<GetBackupPolicyRequest>, ) -> Result<Response<BackupPolicy>, Status>

Returns the description of the specified backup policy by backup_policy_id.

pub async fn list_backup_policies( &mut self, request: impl IntoRequest<ListBackupPoliciesRequest>, ) -> Result<Response<ListBackupPoliciesResponse>, Status>

Returns list of all available backup policies.

pub async fn update_backup_policy( &mut self, request: impl IntoRequest<UpdateBackupPolicyRequest>, ) -> Result<Response<Operation>, Status>

Updates settings of a specific backup policy.

pub async fn delete_backup_policy( &mut self, request: impl IntoRequest<DeleteBackupPolicyRequest>, ) -> Result<Response<Operation>, Status>

Warning! This operation will permanently delete the backup policy.