Struct google_api_proto::google::cloud::netapp::v1::Volume

pub struct Volume {

    pub name: String,
    pub state: i32,
    pub state_details: String,
    pub create_time: Option<Timestamp>,
    pub share_name: String,
    pub psa_range: String,
    pub storage_pool: String,
    pub network: String,
    pub service_level: i32,
    pub capacity_gib: i64,
    pub export_policy: Option<ExportPolicy>,
    pub protocols: Vec<i32>,
    pub smb_settings: Vec<i32>,
    pub mount_options: Vec<MountOption>,
    pub unix_permissions: String,
    pub labels: BTreeMap<String, String>,
    pub description: String,
    pub snapshot_policy: Option<SnapshotPolicy>,
    pub snap_reserve: f64,
    pub snapshot_directory: bool,
    pub used_gib: i64,
    pub security_style: i32,
    pub kerberos_enabled: bool,
    pub ldap_enabled: bool,
    pub active_directory: String,
    pub restore_parameters: Option<RestoreParameters>,
    pub kms_config: String,
    pub encryption_type: i32,
    pub has_replication: bool,
    pub backup_config: Option<BackupConfig>,
    pub restricted_actions: Vec<i32>,

}

Volume provides a filesystem that you can mount.

Fields
name: String

Identifier. Name of the volume

state: i32

Output only. State of the volume

state_details: String

Output only. State details of the volume

create_time: Option<Timestamp>

Output only. Create time of the volume

share_name: String

Required. Share name of the volume

psa_range: String

Output only. Name of the Private Service Access allocated range. This is optional. If not provided, any available range will be chosen.

storage_pool: String

Required. StoragePool name of the volume

network: String

Output only. VPC Network name. Format: projects/{project}/global/networks/{network}

service_level: i32

Output only. Service level of the volume

capacity_gib: i64

Required. Capacity in GIB of the volume

export_policy: Option<ExportPolicy>

Optional. Export policy of the volume

protocols: Vec<i32>

Required. Protocols required for the volume

smb_settings: Vec<i32>

Optional. SMB share settings for the volume.

mount_options: Vec<MountOption>

Output only. Mount options of this volume

unix_permissions: String

Optional. Default unix style permission (e.g. 777) the mount point will be created with. Applicable for NFS protocol types only.

labels: BTreeMap<String, String>

Optional. Labels as key value pairs

description: String

Optional. Description of the volume

snapshot_policy: Option<SnapshotPolicy>

Optional. SnapshotPolicy for a volume.

snap_reserve: f64

Optional. Snap_reserve specifies percentage of volume storage reserved for snapshot storage. Default is 0 percent.

snapshot_directory: bool

Optional. Snapshot_directory if enabled (true) the volume will contain a read-only .snapshot directory which provides access to each of the volume’s snapshots.

used_gib: i64

Output only. Used capacity in GIB of the volume. This is computed periodically and it does not represent the realtime usage.

security_style: i32

Optional. Security Style of the Volume

kerberos_enabled: bool

Optional. Flag indicating if the volume is a kerberos volume or not, export policy rules control kerberos security modes (krb5, krb5i, krb5p).

ldap_enabled: bool

Output only. Flag indicating if the volume is NFS LDAP enabled or not.

active_directory: String

Output only. Specifies the ActiveDirectory name of a SMB volume.

restore_parameters: Option<RestoreParameters>

Optional. Specifies the source of the volume to be created from.

kms_config: String

Output only. Specifies the KMS config to be used for volume encryption.

encryption_type: i32

Output only. Specified the current volume encryption key source.

has_replication: bool

Output only. Indicates whether the volume is part of a replication relationship.

backup_config: Option<BackupConfig>

BackupConfig of the volume.

restricted_actions: Vec<i32>

Optional. List of actions that are restricted on this volume.

Implementations

impl Volume

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.

pub fn service_level(&self) -> ServiceLevel

Returns the enum value of service_level, or the default if the field is set to an invalid enum value.

pub fn set_service_level(&mut self, value: ServiceLevel)

Sets service_level to the provided enum value.

pub fn protocols( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<Protocols>>

Returns an iterator which yields the valid enum values contained in protocols.

pub fn push_protocols(&mut self, value: Protocols)

Appends the provided enum value to protocols.

pub fn smb_settings( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SmbSettings>>

Returns an iterator which yields the valid enum values contained in smb_settings.

pub fn push_smb_settings(&mut self, value: SmbSettings)

Appends the provided enum value to smb_settings.

pub fn security_style(&self) -> SecurityStyle

Returns the enum value of security_style, or the default if the field is set to an invalid enum value.

pub fn set_security_style(&mut self, value: SecurityStyle)

Sets security_style to the provided enum value.

pub fn encryption_type(&self) -> EncryptionType

Returns the enum value of encryption_type, or the default if the field is set to an invalid enum value.

pub fn set_encryption_type(&mut self, value: EncryptionType)

Sets encryption_type to the provided enum value.

pub fn restricted_actions( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<RestrictedAction>>

Returns an iterator which yields the valid enum values contained in restricted_actions.

pub fn push_restricted_actions(&mut self, value: RestrictedAction)

Appends the provided enum value to restricted_actions.