Enum google_api_proto::google::cloud::networkanalyzer::logging::v1::ReportCauseCode

#[repr(i32)]
pub enum ReportCauseCode {

    Unspecified = 0,
    RouteInvalidNextHopVmIpForwardingDisabled = 1,
    RouteInvalidNextHopVmDeleted = 2,
    RouteInvalidNextHopVmStopped = 3,
    RouteInvalidNextHopIlbMisconfigured = 4,
    RouteInvalidNextHopVpnTunnelDeleted = 5,
    RouteInvalidNextHopIlbBackendIpForwardingDisabled = 6,
    IpUtilizationIpAllocationRatioHigh = 20,
    IpUtilizationIpAllocationSummary = 21,
    GkeNodeToControlPlaneBlockedByRoutingIssue = 201,
    GkeNodeToControlPlanePublicEndpointBlockedByEgressFirewall = 202,
    GkeNodeToControlPlanePrivateEndpointBlockedByEgressFirewall = 203,
    GkeControlPlaneToNodeBlockedByRoutingIssue = 211,
    GkeControlPlaneToNodeBlockedByIngressFirewallOnNode = 212,
    GkeIpUtilizationPodRangesAllocationHigh = 221,
    GkeIpUtilizationPodRangesAllocationLimitesAutoscaling = 222,
    GkeNodeServiceAccountServiceAccountDisabled = 243,
    GkeNodeServiceAccountDefaultServiceAccountUsed = 244,
    GkeNodeServiceAccountBadOauthScopes = 245,
    CloudSqlPrivateIpBlockedByEgressFirewall = 601,
    CloudSqlPrivateIpBlockedByRoutingIssue = 602,
    CloudSqlPrivateIpInstanceNotRunning = 603,
    DynamicRouteShadowedFullyShadowedBySubnetRoute = 801,
    DynamicRouteShadowedFullyShadowedByPeeringSubnetRoute = 802,
    DynamicRouteShadowedFullyShadowedByStaticRoute = 803,
    DynamicRouteShadowedFullyShadowedByPeeringStaticRoute = 804,
    DynamicRouteShadowedPartiallyShadowedBySubnetRoute = 805,
    DynamicRouteShadowedPartiallyShadowedByPeeringSubnetRoute = 806,
    DynamicRouteShadowedPartiallyShadowedByStaticRoute = 807,
    DynamicRouteShadowedPartiallyShadowedByPeeringStaticRoute = 808,
    LoadBalancerHealthCheckFirewallHealthCheckFirewallNotConfigured = 1_001,
    LoadBalancerHealthCheckFirewallHealthCheckRangeBlocked = 1_002,
    LoadBalancerHealthCheckFirewallFirewallConfigInconsistent = 1_003,
    LoadBalancerHealthCheckFirewallHealthCheckRangePartiallyBlocked = 1_004,
    LoadBalancerBestPracticesBackendServiceBalancingModeBreaksSessionAffinity = 1_021,
    LoadBalancerBestPracticesBackendServiceHealthCheckPortMismatch = 1_024,

}

Variants

Unspecified = 0

RouteInvalidNextHopVmIpForwardingDisabled = 1

VPC Basics

RouteInvalidNextHopVmDeleted = 2

RouteInvalidNextHopVmStopped = 3

RouteInvalidNextHopIlbMisconfigured = 4

RouteInvalidNextHopVpnTunnelDeleted = 5

RouteInvalidNextHopIlbBackendIpForwardingDisabled = 6

IpUtilizationIpAllocationRatioHigh = 20

IpUtilizationIpAllocationSummary = 21

Summary of ip utilization of all subnet ranges in the project.

GkeNodeToControlPlaneBlockedByRoutingIssue = 201

Kubernetes Engine

GkeNodeToControlPlanePublicEndpointBlockedByEgressFirewall = 202

GkeNodeToControlPlanePrivateEndpointBlockedByEgressFirewall = 203

GkeControlPlaneToNodeBlockedByRoutingIssue = 211

GkeControlPlaneToNodeBlockedByIngressFirewallOnNode = 212

GkeIpUtilizationPodRangesAllocationHigh = 221

GkeIpUtilizationPodRangesAllocationLimitesAutoscaling = 222

GkeNodeServiceAccountServiceAccountDisabled = 243

GkeNodeServiceAccountDefaultServiceAccountUsed = 244

GkeNodeServiceAccountBadOauthScopes = 245

CloudSqlPrivateIpBlockedByEgressFirewall = 601

Managed Services

CloudSqlPrivateIpBlockedByRoutingIssue = 602

CloudSqlPrivateIpInstanceNotRunning = 603

DynamicRouteShadowedFullyShadowedBySubnetRoute = 801

Hybrid Connectivity

DynamicRouteShadowedFullyShadowedByPeeringSubnetRoute = 802

DynamicRouteShadowedFullyShadowedByStaticRoute = 803

DynamicRouteShadowedFullyShadowedByPeeringStaticRoute = 804

DynamicRouteShadowedPartiallyShadowedBySubnetRoute = 805

DynamicRouteShadowedPartiallyShadowedByPeeringSubnetRoute = 806

DynamicRouteShadowedPartiallyShadowedByStaticRoute = 807

DynamicRouteShadowedPartiallyShadowedByPeeringStaticRoute = 808

LoadBalancerHealthCheckFirewallHealthCheckFirewallNotConfigured = 1_001

Network Services

LoadBalancerHealthCheckFirewallHealthCheckRangeBlocked = 1_002

LoadBalancerHealthCheckFirewallFirewallConfigInconsistent = 1_003

LoadBalancerHealthCheckFirewallHealthCheckRangePartiallyBlocked = 1_004

LoadBalancerBestPracticesBackendServiceBalancingModeBreaksSessionAffinity = 1_021

LoadBalancerBestPracticesBackendServiceHealthCheckPortMismatch = 1_024

Implementations

impl ReportCauseCode

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ReportCauseCode.

pub fn from_i32(value: i32) -> Option<ReportCauseCode>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ReportCauseCode, or None if value is not a valid variant.

impl ReportCauseCode

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.