Module google_api_proto::google::cloud::networkconnectivity::v1

Modules

hub_service_client

Generated client implementations.

list_hub_spokes_request

Nested message and enum types in ListHubSpokesRequest.

policy_based_route

Nested message and enum types in PolicyBasedRoute.

policy_based_routing_service_client

Generated client implementations.

spoke

Nested message and enum types in Spoke.

spoke_summary

Nested message and enum types in SpokeSummary.

Structs

AcceptHubSpokeRequest

The request for [HubService.AcceptHubSpoke][google.cloud.networkconnectivity.v1.HubService.AcceptHubSpoke].

AcceptHubSpokeResponse

The response for [HubService.AcceptHubSpoke][google.cloud.networkconnectivity.v1.HubService.AcceptHubSpoke].

CreateHubRequest

Request for [HubService.CreateHub][google.cloud.networkconnectivity.v1.HubService.CreateHub] method.

CreatePolicyBasedRouteRequest

Request for [PolicyBasedRouting.CreatePolicyBasedRoute][] method.

CreateSpokeRequest

The request for [HubService.CreateSpoke][google.cloud.networkconnectivity.v1.HubService.CreateSpoke].

DeleteHubRequest

The request for [HubService.DeleteHub][google.cloud.networkconnectivity.v1.HubService.DeleteHub].

DeletePolicyBasedRouteRequest

Request for [PolicyBasedRouting.DeletePolicyBasedRoute][] method.

DeleteSpokeRequest

The request for [HubService.DeleteSpoke][google.cloud.networkconnectivity.v1.HubService.DeleteSpoke].

GetGroupRequest

The request for [HubService.GetGroup][google.cloud.networkconnectivity.v1.HubService.GetGroup].

GetHubRequest

Request for [HubService.GetHub][google.cloud.networkconnectivity.v1.HubService.GetHub] method.

GetPolicyBasedRouteRequest

Request for [PolicyBasedRouting.GetPolicyBasedRoute][] method.

GetRouteRequest

The request for [HubService.GetRoute][google.cloud.networkconnectivity.v1.HubService.GetRoute].

GetRouteTableRequest

The request for [HubService.GetRouteTable][google.cloud.networkconnectivity.v1.HubService.GetRouteTable].

GetSpokeRequest

The request for [HubService.GetSpoke][google.cloud.networkconnectivity.v1.HubService.GetSpoke].

Group

A group represents a subset of spokes attached to a hub.

Hub

A Network Connectivity Center hub is a global management resource to which you attach spokes. A single hub can contain spokes from multiple regions. However, if any of a hub’s spokes use the site-to-site data transfer feature, the resources associated with those spokes must all be in the same VPC network. Spokes that do not use site-to-site data transfer can be associated with any VPC network in your project.

LinkedInterconnectAttachments

A collection of VLAN attachment resources. These resources should be redundant attachments that all advertise the same prefixes to Google Cloud. Alternatively, in active/passive configurations, all attachments should be capable of advertising the same prefixes.

LinkedRouterApplianceInstances

A collection of router appliance instances. If you configure multiple router appliance instances to receive data from the same set of sites outside of Google Cloud, we recommend that you associate those instances with the same spoke.

LinkedVpcNetwork

An existing VPC network.

LinkedVpnTunnels

A collection of Cloud VPN tunnel resources. These resources should be redundant HA VPN tunnels that all advertise the same prefixes to Google Cloud. Alternatively, in a passive/active configuration, all tunnels should be capable of advertising the same prefixes.

ListGroupsRequest

Request for [HubService.ListGroups][google.cloud.networkconnectivity.v1.HubService.ListGroups] method.

ListGroupsResponse

Response for [HubService.ListGroups][google.cloud.networkconnectivity.v1.HubService.ListGroups] method.

ListHubSpokesRequest

The request for [HubService.ListHubSpokes][google.cloud.networkconnectivity.v1.HubService.ListHubSpokes].

ListHubSpokesResponse

The response for [HubService.ListHubSpokes][google.cloud.networkconnectivity.v1.HubService.ListHubSpokes].

ListHubsRequest

Request for [HubService.ListHubs][google.cloud.networkconnectivity.v1.HubService.ListHubs] method.

ListHubsResponse

Response for [HubService.ListHubs][google.cloud.networkconnectivity.v1.HubService.ListHubs] method.

ListPolicyBasedRoutesRequest

Request for [PolicyBasedRouting.ListPolicyBasedRoutes][] method.

ListPolicyBasedRoutesResponse

Response for [PolicyBasedRouting.ListPolicyBasedRoutes][] method.

ListRouteTablesRequest

Request for [HubService.ListRouteTables][google.cloud.networkconnectivity.v1.HubService.ListRouteTables] method.

ListRouteTablesResponse

Response for [HubService.ListRouteTables][google.cloud.networkconnectivity.v1.HubService.ListRouteTables] method.

ListRoutesRequest

Request for [HubService.ListRoutes][google.cloud.networkconnectivity.v1.HubService.ListRoutes] method.

ListRoutesResponse

Response for [HubService.ListRoutes][google.cloud.networkconnectivity.v1.HubService.ListRoutes] method.

ListSpokesRequest

The request for [HubService.ListSpokes][google.cloud.networkconnectivity.v1.HubService.ListSpokes].

ListSpokesResponse

The response for [HubService.ListSpokes][google.cloud.networkconnectivity.v1.HubService.ListSpokes].

LocationMetadata

Metadata about locations

NextHopVpcNetwork

OperationMetadata

Represents the metadata of the long-running operation.

PolicyBasedRoute

Policy Based Routes (PBR) are more powerful routes that allows GCP customers to route their L4 network traffic based on not just destination IP, but also source IP, protocol and more. A PBR always take precedence when it conflicts with other types of routes. Next id: 22

RejectHubSpokeRequest

The request for [HubService.RejectHubSpoke][google.cloud.networkconnectivity.v1.HubService.RejectHubSpoke].

RejectHubSpokeResponse

The response for [HubService.RejectHubSpoke][google.cloud.networkconnectivity.v1.HubService.RejectHubSpoke].

Route

A route defines a path from VM instances within a spoke to a specific destination resource. Only VPC spokes have routes.

RouteTable

RouterApplianceInstance

A router appliance instance is a Compute Engine virtual machine (VM) instance that acts as a BGP speaker. A router appliance instance is specified by the URI of the VM and the internal IP address of one of the VM’s network interfaces.

RoutingVpc

RoutingVPC contains information about the VPC networks associated with the spokes of a Network Connectivity Center hub.

Spoke

A Network Connectivity Center spoke represents one or more network connectivity resources.

SpokeSummary

Summarizes information about the spokes associated with a hub. The summary includes a count of spokes according to type and according to state. If any spokes are inactive, the summary also lists the reasons they are inactive, including a count for each reason.

UpdateHubRequest

Request for [HubService.UpdateHub][google.cloud.networkconnectivity.v1.HubService.UpdateHub] method.

UpdateSpokeRequest

Request for [HubService.UpdateSpoke][google.cloud.networkconnectivity.v1.HubService.UpdateSpoke] method.

Enums

LocationFeature

Supported features for a location

RouteType

The route’s type

SpokeType

The SpokeType enum represents the type of spoke. The type reflects the kind of resource that a spoke is associated with.

State

The State enum represents the lifecycle stage of a Network Connectivity Center resource.