Module google_api_proto::google::cloud::networkconnectivity::v1::policy_based_route

Nested message and enum types in PolicyBasedRoute.

Modules

filter

Nested message and enum types in Filter.

warnings

Nested message and enum types in Warnings.

Structs

Filter

Filter matches L4 traffic.

InterconnectAttachment

InterconnectAttachment to which this route applies to.

VirtualMachine

VM instances to which this policy based route applies to.

Warnings

Informational warning message.

Enums

NextHop

OtherRoutes

The other routing cases.

Target

Target specifies network endpoints to which this policy based route applies to. If none of the target is specified, the PBR will be installed on all network endpoints (e.g. VMs, VPNs, and Interconnects) in the VPC.