Enum google_api_proto::google::cloud::networkmanagement::v1::drop_info::Cause

#[repr(i32)]
pub enum Cause {

    Unspecified = 0,
    UnknownExternalAddress = 1,
    ForeignIpDisallowed = 2,
    FirewallRule = 3,
    NoRoute = 4,
    RouteBlackhole = 5,
    RouteWrongNetwork = 6,
    RouteNextHopIpAddressNotResolved = 42,
    RouteNextHopResourceNotFound = 43,
    RouteNextHopInstanceWrongNetwork = 49,
    RouteNextHopInstanceNonPrimaryIp = 50,
    RouteNextHopForwardingRuleIpMismatch = 51,
    RouteNextHopVpnTunnelNotEstablished = 52,
    RouteNextHopForwardingRuleTypeInvalid = 53,
    NoRouteFromInternetToPrivateIpv6Address = 44,
    VpnTunnelLocalSelectorMismatch = 45,
    VpnTunnelRemoteSelectorMismatch = 46,
    PrivateTrafficToInternet = 7,
    PrivateGoogleAccessDisallowed = 8,
    PrivateGoogleAccessViaVpnTunnelUnsupported = 47,
    NoExternalAddress = 9,
    UnknownInternalAddress = 10,
    ForwardingRuleMismatch = 11,
    ForwardingRuleNoInstances = 12,
    FirewallBlockingLoadBalancerBackendHealthCheck = 13,
    InstanceNotRunning = 14,
    GkeClusterNotRunning = 27,
    CloudSqlInstanceNotRunning = 28,
    TrafficTypeBlocked = 15,
    GkeMasterUnauthorizedAccess = 16,
    CloudSqlInstanceUnauthorizedAccess = 17,
    DroppedInsideGkeService = 18,
    DroppedInsideCloudSqlService = 19,
    GoogleManagedServiceNoPeering = 20,
    GoogleManagedServiceNoPscEndpoint = 38,
    GkePscEndpointMissing = 36,
    CloudSqlInstanceNoIpAddress = 21,
    GkeControlPlaneRegionMismatch = 30,
    PublicGkeControlPlaneToPrivateDestination = 31,
    GkeControlPlaneNoRoute = 32,
    CloudSqlInstanceNotConfiguredForExternalTraffic = 33,
    PublicCloudSqlInstanceToPrivateDestination = 34,
    CloudSqlInstanceNoRoute = 35,
    CloudFunctionNotActive = 22,
    VpcConnectorNotSet = 23,
    VpcConnectorNotRunning = 24,
    ForwardingRuleRegionMismatch = 25,
    PscConnectionNotAccepted = 26,
    PscEndpointAccessedFromPeeredNetwork = 41,
    PscNegProducerEndpointNoGlobalAccess = 48,
    PscNegProducerForwardingRuleMultiplePorts = 54,
    CloudSqlPscNegUnsupported = 58,
    NoNatSubnetsForPscServiceAttachment = 57,
    HybridNegNonDynamicRouteMatched = 55,
    HybridNegNonLocalDynamicRouteMatched = 56,
    CloudRunRevisionNotReady = 29,
    DroppedInsidePscServiceProducer = 37,
    LoadBalancerHasNoProxySubnet = 39,
    CloudNatNoAddresses = 40,
    RoutingLoop = 59,

}

Drop cause types:

Variants

Unspecified = 0

Cause is unspecified.

UnknownExternalAddress = 1

Destination external address cannot be resolved to a known target. If the address is used in a Google Cloud project, provide the project ID as test input.

ForeignIpDisallowed = 2

A Compute Engine instance can only send or receive a packet with a foreign IP address if ip_forward is enabled.

FirewallRule = 3

Dropped due to a firewall rule, unless allowed due to connection tracking.

NoRoute = 4

Dropped due to no matching routes.

RouteBlackhole = 5

Dropped due to invalid route. Route’s next hop is a blackhole.

RouteWrongNetwork = 6

Packet is sent to a wrong (unintended) network. Example: you trace a packet from VM1:Network1 to VM2:Network2, however, the route configured in Network1 sends the packet destined for VM2’s IP address to Network3.

RouteNextHopIpAddressNotResolved = 42

Route’s next hop IP address cannot be resolved to a GCP resource.

RouteNextHopResourceNotFound = 43

Route’s next hop resource is not found.

RouteNextHopInstanceWrongNetwork = 49

Route’s next hop instance doesn’t have a NIC in the route’s network.

RouteNextHopInstanceNonPrimaryIp = 50

Route’s next hop IP address is not a primary IP address of the next hop instance.

RouteNextHopForwardingRuleIpMismatch = 51

Route’s next hop forwarding rule doesn’t match next hop IP address.

RouteNextHopVpnTunnelNotEstablished = 52

Route’s next hop VPN tunnel is down (does not have valid IKE SAs).

RouteNextHopForwardingRuleTypeInvalid = 53

Route’s next hop forwarding rule type is invalid (it’s not a forwarding rule of the internal passthrough load balancer).

NoRouteFromInternetToPrivateIpv6Address = 44

Packet is sent from the Internet to the private IPv6 address.

VpnTunnelLocalSelectorMismatch = 45

The packet does not match a policy-based VPN tunnel local selector.

VpnTunnelRemoteSelectorMismatch = 46

The packet does not match a policy-based VPN tunnel remote selector.

PrivateTrafficToInternet = 7

Packet with internal destination address sent to the internet gateway.

PrivateGoogleAccessDisallowed = 8

Instance with only an internal IP address tries to access Google API and services, but private Google access is not enabled in the subnet.

PrivateGoogleAccessViaVpnTunnelUnsupported = 47

Source endpoint tries to access Google API and services through the VPN tunnel to another network, but Private Google Access needs to be enabled in the source endpoint network.

NoExternalAddress = 9

Instance with only an internal IP address tries to access external hosts, but Cloud NAT is not enabled in the subnet, unless special configurations on a VM allow this connection.

UnknownInternalAddress = 10

Destination internal address cannot be resolved to a known target. If this is a shared VPC scenario, verify if the service project ID is provided as test input. Otherwise, verify if the IP address is being used in the project.

ForwardingRuleMismatch = 11

Forwarding rule’s protocol and ports do not match the packet header.

ForwardingRuleNoInstances = 12

Forwarding rule does not have backends configured.

FirewallBlockingLoadBalancerBackendHealthCheck = 13

Firewalls block the health check probes to the backends and cause the backends to be unavailable for traffic from the load balancer. For more details, see Health check firewall rules.

InstanceNotRunning = 14

Packet is sent from or to a Compute Engine instance that is not in a running state.

GkeClusterNotRunning = 27

Packet sent from or to a GKE cluster that is not in running state.

CloudSqlInstanceNotRunning = 28

Packet sent from or to a Cloud SQL instance that is not in running state.

TrafficTypeBlocked = 15

The type of traffic is blocked and the user cannot configure a firewall rule to enable it. See Always blocked traffic for more details.

GkeMasterUnauthorizedAccess = 16

Access to Google Kubernetes Engine cluster master’s endpoint is not authorized. See Access to the cluster endpoints for more details.

CloudSqlInstanceUnauthorizedAccess = 17

Access to the Cloud SQL instance endpoint is not authorized. See Authorizing with authorized networks for more details.

DroppedInsideGkeService = 18

Packet was dropped inside Google Kubernetes Engine Service.

DroppedInsideCloudSqlService = 19

Packet was dropped inside Cloud SQL Service.

GoogleManagedServiceNoPeering = 20

Packet was dropped because there is no peering between the originating network and the Google Managed Services Network.

GoogleManagedServiceNoPscEndpoint = 38

Packet was dropped because the Google-managed service uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

GkePscEndpointMissing = 36

Packet was dropped because the GKE cluster uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

CloudSqlInstanceNoIpAddress = 21

Packet was dropped because the Cloud SQL instance has neither a private nor a public IP address.

GkeControlPlaneRegionMismatch = 30

Packet was dropped because a GKE cluster private endpoint is unreachable from a region different from the cluster’s region.

PublicGkeControlPlaneToPrivateDestination = 31

Packet sent from a public GKE cluster control plane to a private IP address.

GkeControlPlaneNoRoute = 32

Packet was dropped because there is no route from a GKE cluster control plane to a destination network.

CloudSqlInstanceNotConfiguredForExternalTraffic = 33

Packet sent from a Cloud SQL instance to an external IP address is not allowed. The Cloud SQL instance is not configured to send packets to external IP addresses.

PublicCloudSqlInstanceToPrivateDestination = 34

Packet sent from a Cloud SQL instance with only a public IP address to a private IP address.

CloudSqlInstanceNoRoute = 35

Packet was dropped because there is no route from a Cloud SQL instance to a destination network.

CloudFunctionNotActive = 22

Packet could be dropped because the Cloud Function is not in an active status.

VpcConnectorNotSet = 23

Packet could be dropped because no VPC connector is set.

VpcConnectorNotRunning = 24

Packet could be dropped because the VPC connector is not in a running state.

ForwardingRuleRegionMismatch = 25

Packet could be dropped because it was sent from a different region to a regional forwarding without global access.

PscConnectionNotAccepted = 26

The Private Service Connect endpoint is in a project that is not approved to connect to the service.

PscEndpointAccessedFromPeeredNetwork = 41

The packet is sent to the Private Service Connect endpoint over the peering, but it’s not supported.

PscNegProducerEndpointNoGlobalAccess = 48

The packet is sent to the Private Service Connect backend (network endpoint group), but the producer PSC forwarding rule does not have global access enabled.

PscNegProducerForwardingRuleMultiplePorts = 54

The packet is sent to the Private Service Connect backend (network endpoint group), but the producer PSC forwarding rule has multiple ports specified.

CloudSqlPscNegUnsupported = 58

The packet is sent to the Private Service Connect backend (network endpoint group) targeting a Cloud SQL service attachment, but this configuration is not supported.

NoNatSubnetsForPscServiceAttachment = 57

No NAT subnets are defined for the PSC service attachment.

HybridNegNonDynamicRouteMatched = 55

The packet sent from the hybrid NEG proxy matches a non-dynamic route, but such a configuration is not supported.

HybridNegNonLocalDynamicRouteMatched = 56

The packet sent from the hybrid NEG proxy matches a dynamic route with a next hop in a different region, but such a configuration is not supported.

CloudRunRevisionNotReady = 29

Packet sent from a Cloud Run revision that is not ready.

DroppedInsidePscServiceProducer = 37

Packet was dropped inside Private Service Connect service producer.

LoadBalancerHasNoProxySubnet = 39

Packet sent to a load balancer, which requires a proxy-only subnet and the subnet is not found.

CloudNatNoAddresses = 40

Packet sent to Cloud Nat without active NAT IPs.

RoutingLoop = 59

Packet is stuck in a routing loop.

Implementations

impl Cause

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Cause.

pub fn from_i32(value: i32) -> Option<Cause>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Cause, or None if value is not a valid variant.

impl Cause

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.