Enum google_api_proto::google::cloud::networkmanagement::v1beta1::drop_info::Cause

#[repr(i32)]
pub enum Cause {

    Unspecified,
    UnknownExternalAddress,
    ForeignIpDisallowed,
    FirewallRule,
    NoRoute,
    RouteBlackhole,
    RouteWrongNetwork,
    PrivateTrafficToInternet,
    PrivateGoogleAccessDisallowed,
    NoExternalAddress,
    UnknownInternalAddress,
    ForwardingRuleMismatch,
    ForwardingRuleRegionMismatch,
    ForwardingRuleNoInstances,
    FirewallBlockingLoadBalancerBackendHealthCheck,
    InstanceNotRunning,
    GkeClusterNotRunning,
    CloudSqlInstanceNotRunning,
    TrafficTypeBlocked,
    GkeMasterUnauthorizedAccess,
    CloudSqlInstanceUnauthorizedAccess,
    DroppedInsideGkeService,
    DroppedInsideCloudSqlService,
    GoogleManagedServiceNoPeering,
    GoogleManagedServiceNoPscEndpoint,
    GkePscEndpointMissing,
    CloudSqlInstanceNoIpAddress,
    GkeControlPlaneRegionMismatch,
    PublicGkeControlPlaneToPrivateDestination,
    GkeControlPlaneNoRoute,
    CloudSqlInstanceNotConfiguredForExternalTraffic,
    PublicCloudSqlInstanceToPrivateDestination,
    CloudSqlInstanceNoRoute,
    CloudFunctionNotActive,
    VpcConnectorNotSet,
    VpcConnectorNotRunning,
    PscConnectionNotAccepted,
    CloudRunRevisionNotReady,
    DroppedInsidePscServiceProducer,
    LoadBalancerHasNoProxySubnet,

}

Drop cause types:

Variants

Unspecified

Cause is unspecified.

UnknownExternalAddress

Destination external address cannot be resolved to a known target. If the address is used in a Google Cloud project, provide the project ID as test input.

ForeignIpDisallowed

A Compute Engine instance can only send or receive a packet with a foreign IP address if ip_forward is enabled.

FirewallRule

Dropped due to a firewall rule, unless allowed due to connection tracking.

NoRoute

Dropped due to no routes.

RouteBlackhole

Dropped due to invalid route. Routeโs next hop is a blackhole.

RouteWrongNetwork

Packet is sent to a wrong (unintended) network. Example: you trace a packet from VM1:Network1 to VM2:Network2, however, the route configured in Network1 sends the packet destined for VM2โs IP addresss to Network3.

PrivateTrafficToInternet

Packet with internal destination address sent to the internet gateway.

PrivateGoogleAccessDisallowed

Instance with only an internal IP address tries to access Google API and services, but private Google access is not enabled.

NoExternalAddress

Instance with only an internal IP address tries to access external hosts, but Cloud NAT is not enabled in the subnet, unless special configurations on a VM allow this connection.

UnknownInternalAddress

Destination internal address cannot be resolved to a known target. If this is a shared VPC scenario, verify if the service project ID is provided as test input. Otherwise, verify if the IP address is being used in the project.

ForwardingRuleMismatch

Forwarding ruleโs protocol and ports do not match the packet header.

ForwardingRuleRegionMismatch

Packet could be dropped because it was sent from a different region to a regional forwarding without global access.

ForwardingRuleNoInstances

Forwarding rule does not have backends configured.

FirewallBlockingLoadBalancerBackendHealthCheck

Firewalls block the health check probes to the backends and cause the backends to be unavailable for traffic from the load balancer. For more details, see Health check firewall rules.

InstanceNotRunning

Packet is sent from or to a Compute Engine instance that is not in a running state.

GkeClusterNotRunning

Packet sent from or to a GKE cluster that is not in running state.

CloudSqlInstanceNotRunning

Packet sent from or to a Cloud SQL instance that is not in running state.

TrafficTypeBlocked

The type of traffic is blocked and the user cannot configure a firewall rule to enable it. See Always blocked traffic for more details.

GkeMasterUnauthorizedAccess

Access to Google Kubernetes Engine cluster masterโs endpoint is not authorized. See Access to the cluster endpoints for more details.

CloudSqlInstanceUnauthorizedAccess

Access to the Cloud SQL instance endpoint is not authorized. See Authorizing with authorized networks for more details.

DroppedInsideGkeService

Packet was dropped inside Google Kubernetes Engine Service.

DroppedInsideCloudSqlService

Packet was dropped inside Cloud SQL Service.

GoogleManagedServiceNoPeering

Packet was dropped because there is no peering between the originating network and the Google Managed Services Network.

GoogleManagedServiceNoPscEndpoint

Packet was dropped because the Google-managed service uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

GkePscEndpointMissing

Packet was dropped because the GKE cluster uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

CloudSqlInstanceNoIpAddress

Packet was dropped because the Cloud SQL instance has neither a private nor a public IP address.

GkeControlPlaneRegionMismatch

Packet was dropped because a GKE cluster private endpoint is unreachable from a region different from the clusterโs region.

PublicGkeControlPlaneToPrivateDestination

Packet sent from a public GKE cluster control plane to a private IP address.

GkeControlPlaneNoRoute

Packet was dropped because there is no route from a GKE cluster control plane to a destination network.

CloudSqlInstanceNotConfiguredForExternalTraffic

Packet sent from a Cloud SQL instance to an external IP address is not allowed. The Cloud SQL instance is not configured to send packets to external IP addresses.

PublicCloudSqlInstanceToPrivateDestination

Packet sent from a Cloud SQL instance with only a public IP address to a private IP address.

CloudSqlInstanceNoRoute

Packet was dropped because there is no route from a Cloud SQL instance to a destination network.

CloudFunctionNotActive

Packet could be dropped because the Cloud Function is not in an active status.

VpcConnectorNotSet

Packet could be dropped because no VPC connector is set.

VpcConnectorNotRunning

Packet could be dropped because the VPC connector is not in a running state.

PscConnectionNotAccepted

The Private Service Connect endpoint is in a project that is not approved to connect to the service.

CloudRunRevisionNotReady

Packet sent from a Cloud Run revision that is not ready.

DroppedInsidePscServiceProducer

Packet was dropped inside Private Service Connect service producer.

LoadBalancerHasNoProxySubnet

Packet sent to a load balancer, which requires a proxy-only subnet and the subnet is not found.

Implementations

impl Cause

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Cause.

pub fn from_i32(value: i32) -> Option<Cause>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Cause, or None if value is not a valid variant.

impl Cause

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.