Module google_api_proto::google::cloud::networkservices::v1::http_route

Nested message and enum types in HttpRoute.

Modules

fault_injection_policy

Nested message and enum types in FaultInjectionPolicy.

header_match

Nested message and enum types in HeaderMatch.

query_parameter_match

Nested message and enum types in QueryParameterMatch.

redirect

Nested message and enum types in Redirect.

route_match

Nested message and enum types in RouteMatch.

Structs

CorsPolicy

The Specification for allowing client side cross-origin requests.

Destination

Specifications of a destination to which the request should be routed to.

FaultInjectionPolicy

The specification for fault injection introduced into traffic to test the resiliency of clients to destination service failure. As part of fault injection, when clients send requests to a destination, delays can be introduced by client proxy on a percentage of requests before sending those requests to the destination service. Similarly requests can be aborted by client proxy for a percentage of requests.

HeaderMatch

Specifies how to select a route rule based on HTTP request headers.

HeaderModifier

The specification for modifying HTTP header in HTTP request and HTTP response.

QueryParameterMatch

Specifications to match a query parameter in the request.

Redirect

The specification for redirecting traffic.

RequestMirrorPolicy

Specifies the policy on how requests are shadowed to a separate mirrored destination service. The proxy does not wait for responses from the shadow service. Prior to sending traffic to the shadow service, the host/authority header is suffixed with -shadow.

RetryPolicy

The specifications for retries.

RouteAction

The specifications for routing traffic and applying associated policies.

RouteMatch

RouteMatch defines specifications used to match requests. If multiple match types are set, this RouteMatch will match if ALL type of matches are matched.

RouteRule

Specifies how to match traffic and how to route traffic when traffic is matched.

UrlRewrite

The specification for modifying the URL of the request, prior to forwarding the request to the destination.