Module google_api_proto::google::cloud::notebooks::v1::instance

Nested message and enum types in Instance.

Modules

disk

Nested message and enum types in Disk.

upgrade_history_entry

Nested message and enum types in UpgradeHistoryEntry.

Structs

AcceleratorConfig

Definition of a hardware accelerator. Note that not all combinations of type and core_count are valid. Check GPUs on Compute Engine to find a valid combination. TPUs are not supported.

Disk

An instance-attached disk resource.

ShieldedInstanceConfig

A set of Shielded Instance options. Check Images using supported Shielded VM features. Not all combinations are valid.

UpgradeHistoryEntry

The entry of VM image upgrade history.

Enums

AcceleratorType

Definition of the types of hardware accelerators that can be used on this instance.

DiskEncryption

Definition of the disk encryption options.

DiskType

Possible disk types for notebook instances.

Environment

Type of the environment; can be one of VM image, or container image.

NicType

The type of vNIC driver. Default should be UNSPECIFIED_NIC_TYPE.

State

The definition of the states of this instance.