Struct google_api_proto::google::cloud::notebooks::v1::VirtualMachineConfig

pub struct VirtualMachineConfig {

    pub zone: String,
    pub machine_type: String,
    pub container_images: Vec<ContainerImage>,
    pub data_disk: Option<LocalDisk>,
    pub encryption_config: Option<EncryptionConfig>,
    pub shielded_instance_config: Option<RuntimeShieldedInstanceConfig>,
    pub accelerator_config: Option<RuntimeAcceleratorConfig>,
    pub network: String,
    pub subnet: String,
    pub internal_ip_only: bool,
    pub tags: Vec<String>,
    pub guest_attributes: BTreeMap<String, String>,
    pub metadata: BTreeMap<String, String>,
    pub labels: BTreeMap<String, String>,
    pub nic_type: i32,
    pub reserved_ip_range: String,
    pub boot_image: Option<BootImage>,

}

The config settings for virtual machine.

Fields
zone: String

Output only. The zone where the virtual machine is located. If using regional request, the notebooks service will pick a location in the corresponding runtime region. On a get request, zone will always be present. Example:

us-central1-b

machine_type: String

Required. The Compute Engine machine type used for runtimes. Short name is valid. Examples:

n1-standard-2

e2-standard-8

container_images: Vec<ContainerImage>

Optional. Use a list of container images to use as Kernels in the notebook instance.

data_disk: Option<LocalDisk>

Required. Data disk option configuration settings.

encryption_config: Option<EncryptionConfig>

Optional. Encryption settings for virtual machine data disk.

shielded_instance_config: Option<RuntimeShieldedInstanceConfig>

Optional. Shielded VM Instance configuration settings.

accelerator_config: Option<RuntimeAcceleratorConfig>

Optional. The Compute Engine accelerator configuration for this runtime.

network: String

Optional. The Compute Engine network to be used for machine communications. Cannot be specified with subnetwork. If neither network nor subnet is specified, the “default” network of the project is used, if it exists.

A full URL or partial URI. Examples:

<https://www.googleapis.com/compute/v1/projects/\[project_id\]/global/networks/default>

projects/\[project_id\]/global/networks/default

Runtimes are managed resources inside Google Infrastructure. Runtimes support the following network configurations:

Google Managed Network (Network & subnet are empty)

Consumer Project VPC (network & subnet are required). Requires configuring Private Service Access.

Shared VPC (network & subnet are required). Requires configuring Private Service Access.

subnet: String

Optional. The Compute Engine subnetwork to be used for machine communications. Cannot be specified with network.

A full URL or partial URI are valid. Examples:

<https://www.googleapis.com/compute/v1/projects/\[project_id\]/regions/us-east1/subnetworks/sub0>

projects/\[project_id\]/regions/us-east1/subnetworks/sub0

internal_ip_only: bool

Optional. If true, runtime will only have internal IP addresses. By default, runtimes are not restricted to internal IP addresses, and will have ephemeral external IP addresses assigned to each vm. This internal_ip_only restriction can only be enabled for subnetwork enabled networks, and all dependencies must be configured to be accessible without external IP addresses.

tags: Vec<String>

Optional. The Compute Engine tags to add to runtime (see Tagging instances).

guest_attributes: BTreeMap<String, String>

Output only. The Compute Engine guest attributes. (see Project and instance guest attributes).

metadata: BTreeMap<String, String>

Optional. The Compute Engine metadata entries to add to virtual machine. (see Project and instance metadata).

labels: BTreeMap<String, String>

Optional. The labels to associate with this runtime. Label keys must contain 1 to 63 characters, and must conform to RFC 1035. Label values may be empty, but, if present, must contain 1 to 63 characters, and must conform to RFC 1035. No more than 32 labels can be associated with a cluster.

nic_type: i32

Optional. The type of vNIC to be used on this interface. This may be gVNIC or VirtioNet.

reserved_ip_range: String

Optional. Reserved IP Range name is used for VPC Peering. The subnetwork allocation will use the range name if it’s assigned.

Example: managed-notebooks-range-c

 PEERING_RANGE_NAME_3=managed-notebooks-range-c
 gcloud compute addresses create $PEERING_RANGE_NAME_3 \
   --global \
   --prefix-length=24 \
   --description="Google Cloud Managed Notebooks Range 24 c" \
   --network=$NETWORK \
   --addresses=192.168.0.0 \
   --purpose=VPC_PEERING

Field value will be: managed-notebooks-range-c

boot_image: Option<BootImage>

Optional. Boot image metadata used for runtime upgradeability.

Implementations

impl VirtualMachineConfig

pub fn nic_type(&self) -> NicType

Returns the enum value of nic_type, or the default if the field is set to an invalid enum value.

pub fn set_nic_type(&mut self, value: NicType)

Sets nic_type to the provided enum value.