Module google_api_proto::google::cloud::notebooks::v1beta1

Modules

environment

Nested message and enum types in Environment.

instance

Nested message and enum types in Instance.

notebook_service_client

Generated client implementations.

reservation_affinity

Nested message and enum types in ReservationAffinity.

vm_image

Nested message and enum types in VmImage.

Structs

ContainerImage

Definition of a container image for starting a notebook instance with the environment installed in a container.

CreateEnvironmentRequest

Request for creating a notebook environment.

CreateInstanceRequest

Request for creating a notebook instance.

DeleteEnvironmentRequest

Request for deleting a notebook environment.

DeleteInstanceRequest

Request for deleting a notebook instance.

Environment

Definition of a software environment that is used to start a notebook instance.

GetEnvironmentRequest

Request for getting a notebook environment.

GetInstanceRequest

Request for getting a notebook instance.

Instance

The definition of a notebook instance.

IsInstanceUpgradeableRequest

Request for checking if a notebook instance is upgradeable.

IsInstanceUpgradeableResponse

Response for checking if a notebook instance is upgradeable.

ListEnvironmentsRequest

Request for listing environments.

ListEnvironmentsResponse

Response for listing environments.

ListInstancesRequest

Request for listing notebook instances.

ListInstancesResponse

Response for listing notebook instances.

OperationMetadata

Represents the metadata of the long-running operation.

RegisterInstanceRequest

Request for registering a notebook instance.

ReportInstanceInfoRequest

Request for notebook instances to report information to Notebooks API.

ReservationAffinity

Reservation Affinity for consuming Zonal reservation.

ResetInstanceRequest

Request for reseting a notebook instance

SetInstanceAcceleratorRequest

Request for setting instance accelerator.

SetInstanceLabelsRequest

Request for setting instance labels.

SetInstanceMachineTypeRequest

Request for setting instance machine type.

StartInstanceRequest

Request for starting a notebook instance

StopInstanceRequest

Request for stopping a notebook instance

UpgradeInstanceInternalRequest

Request for upgrading a notebook instance from within the VM

UpgradeInstanceRequest

Request for upgrading a notebook instance

VmImage

Definition of a custom Compute Engine virtual machine image for starting a notebook instance with the environment installed directly on the VM.