Module google_api_proto::google::cloud::notebooks::v2

Modules

accelerator_config

Nested message and enum types in AcceleratorConfig.

event

Nested message and enum types in Event.

gce_setup

Nested message and enum types in GceSetup.

instance

Nested message and enum types in Instance.

network_interface

Nested message and enum types in NetworkInterface.

notebook_service_client

Generated client implementations.

upgrade_history_entry

Nested message and enum types in UpgradeHistoryEntry.

vm_image

Nested message and enum types in VmImage.

Structs

AcceleratorConfig

An accelerator configuration for a VM instance Definition of a hardware accelerator. Note that there is no check on type and core_count combinations. TPUs are not supported. See GPUs on Compute Engine to find a valid combination.

BootDisk

The definition of a boot disk.

CheckInstanceUpgradabilityRequest

Request for checking if a notebook instance is upgradeable.

CheckInstanceUpgradabilityResponse

Response for checking if a notebook instance is upgradeable.

ContainerImage

Definition of a container image for starting a notebook instance with the environment installed in a container.

CreateInstanceRequest

Request for creating a notebook instance.

DataDisk

An instance-attached disk resource.

DeleteInstanceRequest

Request for deleting a notebook instance.

DiagnoseInstanceRequest

Request for creating a notebook instance diagnostic file.

DiagnosticConfig

Defines flags that are used to run the diagnostic tool

Event

The definition of an Event for a managed / semi-managed notebook instance.

GceSetup

The definition of how to configure a VM instance outside of Resources and Identity.

GetInstanceRequest

Request for getting a notebook instance.

GpuDriverConfig

A GPU driver configuration

Instance

The definition of a notebook instance.

ListInstancesRequest

Request for listing notebook instances.

ListInstancesResponse

Response for listing notebook instances.

NetworkInterface

The definition of a network interface resource attached to a VM.

OperationMetadata

Represents the metadata of the long-running operation.

ResetInstanceRequest

Request for resetting a notebook instance

RollbackInstanceRequest

Request for rollbacking a notebook instance

ServiceAccount

A service account that acts as an identity.

ShieldedInstanceConfig

A set of Shielded Instance options. See Images using supported Shielded VM features. Not all combinations are valid.

StartInstanceRequest

Request for starting a notebook instance

StopInstanceRequest

Request for stopping a notebook instance

UpdateInstanceRequest

Request for updating a notebook instance.

UpgradeHistoryEntry

The entry of VM image upgrade history.

UpgradeInstanceRequest

Request for upgrading a notebook instance

VmImage

Definition of a custom Compute Engine virtual machine image for starting a notebook instance with the environment installed directly on the VM.

Enums

DiskEncryption

Definition of the disk encryption options.

DiskType

Possible disk types.

HealthState

The instance health state.

State

The definition of the states of this instance.