Module google_api_proto::google::cloud::optimization::v1::break_rule

Nested message and enum types in BreakRule.

Structs

BreakRequest

The sequence of breaks (i.e. their number and order) that apply to each vehicle must be known beforehand. The repeated BreakRequests define that sequence, in the order in which they must occur. Their time windows (earliest_start_time / latest_start_time) may overlap, but they must be compatible with the order (this is checked).

FrequencyConstraint

One may further constrain the frequency and duration of the breaks specified above, by enforcing a minimum break frequency, such as “There must be a break of at least 1 hour every 12 hours”. Assuming that this can be interpreted as “Within any sliding time window of 12h, there must be at least one break of at least one hour”, that example would translate to the following FrequencyConstraint: