Module google_api_proto::google::cloud::optimization::v1::shipment_route

Nested message and enum types in ShipmentRoute.

Structs

Break

Data representing the execution of a break.

Delay

Deprecated: Use [ShipmentRoute.Transition.delay_duration][google.cloud.optimization.v1.ShipmentRoute.Transition.delay_duration] instead. Time interval spent on the route resulting from a [TransitionAttributes.delay][google.cloud.optimization.v1.TransitionAttributes.delay].

EncodedPolyline

The encoded representation of a polyline. More information on polyline encoding can be found here: https://developers.google.com/maps/documentation/utilities/polylinealgorithm https://developers.google.com/maps/documentation/javascript/reference/geometry#encoding.

Transition

Transition between two events on the route. See the description of [ShipmentRoute][google.cloud.optimization.v1.ShipmentRoute].

TravelStep

Deprecated: Use [ShipmentRoute.Transition][google.cloud.optimization.v1.ShipmentRoute.Transition] instead. Travel between each visit along the route: from the vehicle’s start_location to the first visit’s arrival_location, then from the first visit’s departure_location to the second visit’s arrival_location, and so on until the vehicle’s end_location. This accounts only for the actual travel between visits, not counting the waiting time, the time spent performing a visit, nor the distance covered during a visit.

VehicleLoad

Reports the actual load of the vehicle at some point along the route, for a given type (see [Transition.vehicle_loads][google.cloud.optimization.v1.ShipmentRoute.Transition.vehicle_loads]).

Visit

A visit performed during a route. This visit corresponds to a pickup or a delivery of a Shipment.