Struct google_api_proto::google::cloud::orchestration::airflow::service::v1::SoftwareConfig

pub struct SoftwareConfig {
    pub image_version: String,
    pub airflow_config_overrides: BTreeMap<String, String>,
    pub pypi_packages: BTreeMap<String, String>,
    pub env_variables: BTreeMap<String, String>,
    pub python_version: String,
    pub scheduler_count: i32,
    pub cloud_data_lineage_integration: Option<CloudDataLineageIntegration>,
    pub web_server_plugins_mode: i32,
}

Specifies the selection and configuration of software inside the environment.

Fields
image_version: String

The version of the software running in the environment. This encapsulates both the version of Cloud Composer functionality and the version of Apache Airflow. It must match the regular expression composer-(\[0-9\]+(\.\[0-9\]+\.\[0-9\]+(-preview\.\[0-9\]+)?)?|latest)-airflow-(\[0-9\]+(\.\[0-9\]+(\.\[0-9\]+)?)?). When used as input, the server also checks if the provided version is supported and denies the request for an unsupported version.

The Cloud Composer portion of the image version is a full semantic version, or an alias in the form of major version number or latest. When an alias is provided, the server replaces it with the current Cloud Composer version that satisfies the alias.

The Apache Airflow portion of the image version is a full semantic version that points to one of the supported Apache Airflow versions, or an alias in the form of only major or major.minor versions specified. When an alias is provided, the server replaces it with the latest Apache Airflow version that satisfies the alias and is supported in the given Cloud Composer version.

In all cases, the resolved image version is stored in the same field.

See also version list and versioning overview.

airflow_config_overrides: BTreeMap<String, String>

Optional. Apache Airflow configuration properties to override.

Property keys contain the section and property names, separated by a hyphen, for example “core-dags_are_paused_at_creation”. Section names must not contain hyphens (“-”), opening square brackets (“[”), or closing square brackets (“]”). The property name must not be empty and must not contain an equals sign (“=”) or semicolon (“;”). Section and property names must not contain a period (“.”). Apache Airflow configuration property names must be written in snake_case. Property values can contain any character, and can be written in any lower/upper case format.

Certain Apache Airflow configuration property values are blocked, and cannot be overridden.

pypi_packages: BTreeMap<String, String>

Optional. Custom Python Package Index (PyPI) packages to be installed in the environment.

Keys refer to the lowercase package name such as “numpy” and values are the lowercase extras and version specifier such as “==1.12.0”, “[devel,gcp_api]”, or “[devel]>=1.8.2, <1.9.2”. To specify a package without pinning it to a version specifier, use the empty string as the value.

env_variables: BTreeMap<String, String>

Optional. Additional environment variables to provide to the Apache Airflow scheduler, worker, and webserver processes.

Environment variable names must match the regular expression [a-zA-Z_][a-zA-Z0-9_]*. They cannot specify Apache Airflow software configuration overrides (they cannot match the regular expression AIRFLOW__\[A-Z0-9_\]+__\[A-Z0-9_\]+), and they cannot match any of the following reserved names:

AIRFLOW_HOME

C_FORCE_ROOT

CONTAINER_NAME

DAGS_FOLDER

GCP_PROJECT

GCS_BUCKET

GKE_CLUSTER_NAME

SQL_DATABASE

SQL_INSTANCE

SQL_PASSWORD

SQL_PROJECT

SQL_REGION

SQL_USER

python_version: String

Optional. The major version of Python used to run the Apache Airflow scheduler, worker, and webserver processes.

Can be set to ‘2’ or ‘3’. If not specified, the default is ‘3’. Cannot be updated.

This field is only supported for Cloud Composer environments in versions composer-1..-airflow-..*. Environments in newer versions always use Python major version 3.

scheduler_count: i32

Optional. The number of schedulers for Airflow.

This field is supported for Cloud Composer environments in versions composer-1..-airflow-2...

cloud_data_lineage_integration: Option<CloudDataLineageIntegration>

Optional. The configuration for Cloud Data Lineage integration.

web_server_plugins_mode: i32

Optional. Whether or not the web server uses custom plugins. If unspecified, the field defaults to PLUGINS_ENABLED.

This field is supported for Cloud Composer environments in versions composer-3..-airflow-..* and newer.

Implementations

impl SoftwareConfig

pub fn web_server_plugins_mode(&self) -> WebServerPluginsMode

Returns the enum value of web_server_plugins_mode, or the default if the field is set to an invalid enum value.

pub fn set_web_server_plugins_mode(&mut self, value: WebServerPluginsMode)

Sets web_server_plugins_mode to the provided enum value.