Struct google_api_proto::google::cloud::orchestration::airflow::service::v1beta1::EnvironmentConfig

pub struct EnvironmentConfig {

    pub gke_cluster: String,
    pub dag_gcs_prefix: String,
    pub node_count: i32,
    pub software_config: Option<SoftwareConfig>,
    pub node_config: Option<NodeConfig>,
    pub private_environment_config: Option<PrivateEnvironmentConfig>,
    pub web_server_network_access_control: Option<WebServerNetworkAccessControl>,
    pub database_config: Option<DatabaseConfig>,
    pub web_server_config: Option<WebServerConfig>,
    pub airflow_uri: String,
    pub airflow_byoid_uri: String,
    pub encryption_config: Option<EncryptionConfig>,
    pub maintenance_window: Option<MaintenanceWindow>,
    pub workloads_config: Option<WorkloadsConfig>,
    pub environment_size: i32,
    pub master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>,
    pub recovery_config: Option<RecoveryConfig>,
    pub resilience_mode: i32,

}

Configuration information for an environment.

Fields
gke_cluster: String

Output only. The Kubernetes Engine cluster used to run this environment.

dag_gcs_prefix: String

Output only. The Cloud Storage prefix of the DAGs for this environment. Although Cloud Storage objects reside in a flat namespace, a hierarchical file tree can be simulated using “/”-delimited object name prefixes. DAG objects for this environment reside in a simulated directory with the given prefix.

node_count: i32

The number of nodes in the Kubernetes Engine cluster that will be used to run this environment.

This field is supported for Cloud Composer environments in versions composer-1..-airflow-..*.

software_config: Option<SoftwareConfig>

The configuration settings for software inside the environment.

node_config: Option<NodeConfig>

The configuration used for the Kubernetes Engine cluster.

private_environment_config: Option<PrivateEnvironmentConfig>

The configuration used for the Private IP Cloud Composer environment.

web_server_network_access_control: Option<WebServerNetworkAccessControl>

Optional. The network-level access control policy for the Airflow web server. If unspecified, no network-level access restrictions will be applied.

database_config: Option<DatabaseConfig>

Optional. The configuration settings for Cloud SQL instance used internally by Apache Airflow software.

web_server_config: Option<WebServerConfig>

Optional. The configuration settings for the Airflow web server App Engine instance.

This field is supported for Cloud Composer environments in versions composer-1..-airflow-..*.

airflow_uri: String

Output only. The URI of the Apache Airflow Web UI hosted within this environment (see Airflow web interface).

airflow_byoid_uri: String

Output only. The ‘bring your own identity’ variant of the URI of the Apache Airflow Web UI hosted within this environment, to be accessed with external identities using workforce identity federation (see Access environments with workforce identity federation).

encryption_config: Option<EncryptionConfig>

Optional. The encryption options for the Cloud Composer environment and its dependencies. Cannot be updated.

maintenance_window: Option<MaintenanceWindow>

Optional. The maintenance window is the period when Cloud Composer components may undergo maintenance. It is defined so that maintenance is not executed during peak hours or critical time periods.

The system will not be under maintenance for every occurrence of this window, but when maintenance is planned, it will be scheduled during the window.

The maintenance window period must encompass at least 12 hours per week. This may be split into multiple chunks, each with a size of at least 4 hours.

If this value is omitted, Cloud Composer components may be subject to maintenance at any time.

workloads_config: Option<WorkloadsConfig>

Optional. The workloads configuration settings for the GKE cluster associated with the Cloud Composer environment. The GKE cluster runs Airflow scheduler, web server and workers workloads.

This field is supported for Cloud Composer environments in versions composer-2..-airflow-..* and newer.

environment_size: i32

Optional. The size of the Cloud Composer environment.

This field is supported for Cloud Composer environments in versions composer-2..-airflow-..* and newer.

master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>

Optional. The configuration options for GKE cluster master authorized networks. By default master authorized networks feature is:

in case of private environment: enabled with no external networks allowlisted.

in case of public environment: disabled.

recovery_config: Option<RecoveryConfig>

Optional. The Recovery settings configuration of an environment.

This field is supported for Cloud Composer environments in versions composer-2..-airflow-..* and newer.

resilience_mode: i32

Optional. Resilience mode of the Cloud Composer Environment.

This field is supported for Cloud Composer environments in versions composer-2.2.0-airflow-..* and newer.

Implementations

impl EnvironmentConfig

pub fn environment_size(&self) -> EnvironmentSize

Returns the enum value of environment_size, or the default if the field is set to an invalid enum value.

pub fn set_environment_size(&mut self, value: EnvironmentSize)

Sets environment_size to the provided enum value.

pub fn resilience_mode(&self) -> ResilienceMode

Returns the enum value of resilience_mode, or the default if the field is set to an invalid enum value.

pub fn set_resilience_mode(&mut self, value: ResilienceMode)

Sets resilience_mode to the provided enum value.