Struct google_api_proto::google::cloud::orgpolicy::v2::org_policy_client::OrgPolicyClient

pub struct OrgPolicyClient<T> { /* private fields */ }

An interface for managing organization policies.

The Organization Policy Service provides a simple mechanism for organizations to restrict the allowed configurations across their entire resource hierarchy.

You can use a policy to configure restrictions on resources. For example, you can enforce a policy that restricts which Google Cloud APIs can be activated in a certain part of your resource hierarchy, or prevents serial port access to VM instances in a particular folder.

Policies are inherited down through the resource hierarchy. A policy applied to a parent resource automatically applies to all its child resources unless overridden with a policy lower in the hierarchy.

A constraint defines an aspect of a resource’s configuration that can be controlled by an organization’s policy administrator. Policies are a collection of constraints that defines their allowable configuration on a particular resource and its child resources.

Implementations

impl<T> OrgPolicyClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> OrgPolicyClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_constraints( &mut self, request: impl IntoRequest<ListConstraintsRequest>, ) -> Result<Response<ListConstraintsResponse>, Status>

Lists constraints that could be applied on the specified resource.

pub async fn list_policies( &mut self, request: impl IntoRequest<ListPoliciesRequest>, ) -> Result<Response<ListPoliciesResponse>, Status>

Retrieves all of the policies that exist on a particular resource.

pub async fn get_policy( &mut self, request: impl IntoRequest<GetPolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets a policy on a resource.

If no policy is set on the resource, NOT_FOUND is returned. The etag value can be used with UpdatePolicy() to update a policy during read-modify-write.

pub async fn get_effective_policy( &mut self, request: impl IntoRequest<GetEffectivePolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets the effective policy on a resource. This is the result of merging policies in the resource hierarchy and evaluating conditions. The returned policy will not have an etag or condition set because it is an evaluated policy across multiple resources. Subtrees of Resource Manager resource hierarchy with ‘under:’ prefix will not be expanded.

pub async fn create_policy( &mut self, request: impl IntoRequest<CreatePolicyRequest>, ) -> Result<Response<Policy>, Status>

Creates a policy.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist. Returns a google.rpc.Status with google.rpc.Code.ALREADY_EXISTS if the policy already exists on the given Google Cloud resource.

pub async fn update_policy( &mut self, request: impl IntoRequest<UpdatePolicyRequest>, ) -> Result<Response<Policy>, Status>

Updates a policy.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or the policy do not exist. Returns a google.rpc.Status with google.rpc.Code.ABORTED if the etag supplied in the request does not match the persisted etag of the policy

Note: the supplied policy will perform a full overwrite of all fields.

pub async fn delete_policy( &mut self, request: impl IntoRequest<DeletePolicyRequest>, ) -> Result<Response<()>, Status>

Deletes a policy.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint or organization policy does not exist.

pub async fn create_custom_constraint( &mut self, request: impl IntoRequest<CreateCustomConstraintRequest>, ) -> Result<Response<CustomConstraint>, Status>

Creates a custom constraint.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the organization does not exist. Returns a google.rpc.Status with google.rpc.Code.ALREADY_EXISTS if the constraint already exists on the given organization.

pub async fn update_custom_constraint( &mut self, request: impl IntoRequest<UpdateCustomConstraintRequest>, ) -> Result<Response<CustomConstraint>, Status>

Updates a custom constraint.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist.

Note: the supplied policy will perform a full overwrite of all fields.

pub async fn get_custom_constraint( &mut self, request: impl IntoRequest<GetCustomConstraintRequest>, ) -> Result<Response<CustomConstraint>, Status>

Gets a custom constraint.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the custom constraint does not exist.

pub async fn list_custom_constraints( &mut self, request: impl IntoRequest<ListCustomConstraintsRequest>, ) -> Result<Response<ListCustomConstraintsResponse>, Status>

Retrieves all of the custom constraints that exist on a particular organization resource.

pub async fn delete_custom_constraint( &mut self, request: impl IntoRequest<DeleteCustomConstraintRequest>, ) -> Result<Response<()>, Status>

Deletes a custom constraint.

Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the constraint does not exist.