Module google_api_proto::google::cloud::osconfig::v1alpha

Modules

cvs_sv3

Nested message and enum types in CVSSv3.

fixed_or_percent

Nested message and enum types in FixedOrPercent.

instance_os_policies_compliance

Nested message and enum types in InstanceOSPoliciesCompliance.

inventory

Nested message and enum types in Inventory.

os_config_zonal_service_client

Generated client implementations.

os_policy

Nested message and enum types in OSPolicy.

os_policy_assignment

Nested message and enum types in OSPolicyAssignment.

os_policy_assignment_operation_metadata

Nested message and enum types in OSPolicyAssignmentOperationMetadata.

os_policy_assignment_report

Nested message and enum types in OSPolicyAssignmentReport.

os_policy_resource_compliance

Nested message and enum types in OSPolicyResourceCompliance.

os_policy_resource_config_step

Nested message and enum types in OSPolicyResourceConfigStep.

vulnerability_report

Nested message and enum types in VulnerabilityReport.

Structs

CreateOsPolicyAssignmentRequest

A request message to create an OS policy assignment

CvsSv3

Common Vulnerability Scoring System version 3. For details, see https://www.first.org/cvss/specification-document

DeleteOsPolicyAssignmentRequest

A request message for deleting a OS policy assignment.

FixedOrPercent

Message encapsulating a value that can be either absolute (“fixed”) or relative (“percent”) to a value.

GetInstanceOsPoliciesComplianceRequest

A request message for getting OS policies compliance data for the given Compute Engine VM instance.

GetInventoryRequest

A request message for getting inventory data for the specified VM.

GetOsPolicyAssignmentReportRequest

Get a report of the OS policy assignment for a VM instance.

GetOsPolicyAssignmentRequest

A request message to get an OS policy assignment

GetVulnerabilityReportRequest

A request message for getting the vulnerability report for the specified VM.

InstanceOsPoliciesCompliance

This API resource represents the OS policies compliance data for a Compute Engine virtual machine (VM) instance at a given point in time.

Inventory

This API resource represents the available inventory data for a Compute Engine virtual machine (VM) instance at a given point in time.

ListInstanceOsPoliciesCompliancesRequest

A request message for listing OS policies compliance data for all Compute Engine VMs in the given location.

ListInstanceOsPoliciesCompliancesResponse

A response message for listing OS policies compliance data for all Compute Engine VMs in the given location.

ListInventoriesRequest

A request message for listing inventory data for all VMs in the specified location.

ListInventoriesResponse

A response message for listing inventory data for all VMs in a specified location.

ListOsPolicyAssignmentReportsRequest

List the OS policy assignment reports for VM instances.

ListOsPolicyAssignmentReportsResponse

A response message for listing OS Policy assignment reports including the page of results and page token.

ListOsPolicyAssignmentRevisionsRequest

A request message to list revisions for a OS policy assignment

ListOsPolicyAssignmentRevisionsResponse

A response message for listing all revisions for a OS policy assignment.

ListOsPolicyAssignmentsRequest

A request message to list OS policy assignments for a parent resource

ListOsPolicyAssignmentsResponse

A response message for listing all assignments under given parent.

ListVulnerabilityReportsRequest

A request message for listing vulnerability reports for all VM instances in the specified location.

ListVulnerabilityReportsResponse

A response message for listing vulnerability reports for all VM instances in the specified location.

OsPolicy

An OS policy defines the desired state configuration for a VM.

OsPolicyAssignment

OS policy assignment is an API resource that is used to apply a set of OS policies to a dynamically targeted group of Compute Engine VM instances.

OsPolicyAssignmentOperationMetadata

OS policy assignment operation metadata provided by OS policy assignment API methods that return long running operations.

OsPolicyAssignmentReport

A report of the OS policy assignment status for a given instance.

OsPolicyResourceCompliance

Compliance data for an OS policy resource.

OsPolicyResourceConfigStep

Step performed by the OS Config agent for configuring an OSPolicyResource to its desired state.

UpdateOsPolicyAssignmentRequest

A request message to update an OS policy assignment

VulnerabilityReport

This API resource represents the vulnerability report for a specified Compute Engine virtual machine (VM) instance at a given point in time.

Enums

InventoryView

The view for inventory objects.

OsPolicyComplianceState

Supported OSPolicy compliance states.