Struct google_api_proto::google::cloud::osconfig::v1beta::os_config_service_client::OsConfigServiceClient

pub struct OsConfigServiceClient<T> { /* private fields */ }

OS Config API

The OS Config service is a server-side component that you can use to manage package installations and patch jobs for virtual machine instances.

Implementations

impl<T> OsConfigServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> OsConfigServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn execute_patch_job( &mut self, request: impl IntoRequest<ExecutePatchJobRequest> ) -> Result<Response<PatchJob>, Status>

Patch VM instances by creating and running a patch job.

pub async fn get_patch_job( &mut self, request: impl IntoRequest<GetPatchJobRequest> ) -> Result<Response<PatchJob>, Status>

Get the patch job. This can be used to track the progress of an ongoing patch job or review the details of completed jobs.

pub async fn cancel_patch_job( &mut self, request: impl IntoRequest<CancelPatchJobRequest> ) -> Result<Response<PatchJob>, Status>

Cancel a patch job. The patch job must be active. Canceled patch jobs cannot be restarted.

pub async fn list_patch_jobs( &mut self, request: impl IntoRequest<ListPatchJobsRequest> ) -> Result<Response<ListPatchJobsResponse>, Status>

Get a list of patch jobs.

pub async fn list_patch_job_instance_details( &mut self, request: impl IntoRequest<ListPatchJobInstanceDetailsRequest> ) -> Result<Response<ListPatchJobInstanceDetailsResponse>, Status>

Get a list of instance details for a given patch job.

pub async fn create_patch_deployment( &mut self, request: impl IntoRequest<CreatePatchDeploymentRequest> ) -> Result<Response<PatchDeployment>, Status>

Create an OS Config patch deployment.

pub async fn get_patch_deployment( &mut self, request: impl IntoRequest<GetPatchDeploymentRequest> ) -> Result<Response<PatchDeployment>, Status>

Get an OS Config patch deployment.

pub async fn list_patch_deployments( &mut self, request: impl IntoRequest<ListPatchDeploymentsRequest> ) -> Result<Response<ListPatchDeploymentsResponse>, Status>

Get a page of OS Config patch deployments.

pub async fn delete_patch_deployment( &mut self, request: impl IntoRequest<DeletePatchDeploymentRequest> ) -> Result<Response<()>, Status>

Delete an OS Config patch deployment.

pub async fn update_patch_deployment( &mut self, request: impl IntoRequest<UpdatePatchDeploymentRequest> ) -> Result<Response<PatchDeployment>, Status>

Update an OS Config patch deployment.

pub async fn pause_patch_deployment( &mut self, request: impl IntoRequest<PausePatchDeploymentRequest> ) -> Result<Response<PatchDeployment>, Status>

Change state of patch deployment to “PAUSED”. Patch deployment in paused state doesn’t generate patch jobs.

pub async fn resume_patch_deployment( &mut self, request: impl IntoRequest<ResumePatchDeploymentRequest> ) -> Result<Response<PatchDeployment>, Status>

Change state of patch deployment back to “ACTIVE”. Patch deployment in active state continues to generate patch jobs.

pub async fn create_guest_policy( &mut self, request: impl IntoRequest<CreateGuestPolicyRequest> ) -> Result<Response<GuestPolicy>, Status>

Create an OS Config guest policy.

pub async fn get_guest_policy( &mut self, request: impl IntoRequest<GetGuestPolicyRequest> ) -> Result<Response<GuestPolicy>, Status>

Get an OS Config guest policy.

pub async fn list_guest_policies( &mut self, request: impl IntoRequest<ListGuestPoliciesRequest> ) -> Result<Response<ListGuestPoliciesResponse>, Status>

Get a page of OS Config guest policies.

pub async fn update_guest_policy( &mut self, request: impl IntoRequest<UpdateGuestPolicyRequest> ) -> Result<Response<GuestPolicy>, Status>

Update an OS Config guest policy.

pub async fn delete_guest_policy( &mut self, request: impl IntoRequest<DeleteGuestPolicyRequest> ) -> Result<Response<()>, Status>

Delete an OS Config guest policy.

pub async fn lookup_effective_guest_policy( &mut self, request: impl IntoRequest<LookupEffectiveGuestPolicyRequest> ) -> Result<Response<EffectiveGuestPolicy>, Status>

Lookup the effective guest policy that applies to a VM instance. This lookup merges all policies that are assigned to the instance ancestry.