Module google_api_proto::google::cloud::parallelstore::v1beta

Modules

export_data_request

Nested message and enum types in ExportDataRequest.

import_data_request

Nested message and enum types in ImportDataRequest.

instance

Nested message and enum types in Instance.

parallelstore_client

Generated client implementations.

transfer_operation_metadata

Nested message and enum types in TransferOperationMetadata.

Structs

CreateInstanceRequest

Request for [CreateInstance][google.cloud.parallelstore.v1beta.Parallelstore.CreateInstance]

DeleteInstanceRequest

Message for deleting a Instance

DestinationGcsBucket

Google Cloud Storage as a destination.

DestinationParallelstore

Parallelstore as a destination.

ExportDataMetadata

ExportDataMetadata contains export data operation metadata

ExportDataRequest

Message representing the request exporting data from Cloud Storage to parallelstore.

ExportDataResponse

ExportDataResponse is the response returned from ExportData rpc

GetInstanceRequest

Request to get an instance’s details.

ImportDataMetadata

ImportDataMetadata contains import data operation metadata

ImportDataRequest

Message representing the request importing data from parallelstore to Cloud Storage.

ImportDataResponse

ImportDataResponse is the response returned from ImportData rpc.

Instance

A Parallelstore instance.

ListInstancesRequest

Message for requesting list of Instances

ListInstancesResponse

Message for response to listing Instances

OperationMetadata

Represents the metadata of the long-running operation.

SourceGcsBucket

Google Cloud Storage as a source.

SourceParallelstore

Pa as a source.

TransferCounters

A collection of counters that report the progress of a transfer operation.

TransferOperationMetadata

Represents the metadata of the long-running operation.

UpdateInstanceRequest

Message for updating a Instance

Enums

DirectoryStripeLevel

Represents the striping options for directories.

FileStripeLevel

Represents the striping options for files.

TransferType

Type of transfer that occurred.