Module google_api_proto::google::cloud::paymentgateway::issuerswitch::v1

Modules

case_response

Nested message and enum types in CaseResponse.

complaint_transaction

Nested message and enum types in ComplaintTransaction.

financial_transaction

Nested message and enum types in FinancialTransaction.

issuer_participant

Nested message and enum types in IssuerParticipant.

issuer_switch_participants_client

Generated client implementations.

issuer_switch_resolutions_client

Generated client implementations.

issuer_switch_rules_client

Generated client implementations.

issuer_switch_transactions_client

Generated client implementations.

mandate_transaction

Nested message and enum types in MandateTransaction.

merchant_additional_info

Nested message and enum types in MerchantAdditionalInfo.

participant

Nested message and enum types in Participant.

participant_state_change_request

Nested message and enum types in ParticipantStateChangeRequest.

raise_complaint_adjustment

Nested message and enum types in RaiseComplaintAdjustment.

raise_dispute_adjustment

Nested message and enum types in RaiseDisputeAdjustment.

resolve_complaint_adjustment

Nested message and enum types in ResolveComplaintAdjustment.

resolve_dispute_adjustment

Nested message and enum types in ResolveDisputeAdjustment.

rule_metadata

Nested message and enum types in RuleMetadata.

rule_metadata_value

Nested message and enum types in RuleMetadataValue.

settlement_participant

Nested message and enum types in SettlementParticipant.

transaction_info

Nested message and enum types in TransactionInfo.

upi_transaction

Nested message and enum types in UpiTransaction.

Structs

AccountReference

A reference to uniquely identify an account according to India’s UPI standards.

BatchCreateRuleMetadataValuesRequest

Request body for the BatchCreateRuleMetadataValues method.

BatchCreateRuleMetadataValuesResponse

Response body for the BatchCreateRuleMetadataValues method.

BatchDeleteRuleMetadataValuesRequest

Request body for the BatchDeleteRuleMetadataValues method.

CaseDetails

Details of the complaint or dispute.

CaseResponse

Response to the complaint or dispute.

Complaint

A complaint processed by the issuer switch.

ComplaintTransaction

A complaint API transaction processed by the issuer switch. In UPI, this maps to the Complaint API.

CreateComplaintMetadata

Metadata for CreateComplaint.

CreateComplaintRequest

Request for the CreateComplaint method.

CreateDisputeMetadata

Metadata for CreateDispute.

CreateDisputeRequest

Request for the CreateDispute method.

CreateRuleMetadataValueRequest

Request for creating a single RuleMetadataValue.

DeviceDetails

A participant’s device tags

Dispute

A dispute processed by the issuer switch.

ExportComplaintTransactionsMetadata

Metadata for ExportComplaintTransactions.

ExportComplaintTransactionsRequest

Request for the ExportComplaintTransactions method.

ExportComplaintTransactionsResponse

Response for the ExportComplaintTransactions method.

ExportFinancialTransactionsMetadata

Metadata for ExportFinancialTransactions.

ExportFinancialTransactionsRequest

Request for the ExportFinancialTransactions method.

ExportFinancialTransactionsResponse

Response for the ExportFinancialTransactions method.

ExportMandateTransactionsMetadata

Metadata for ExportMandateTransactions.

ExportMandateTransactionsRequest

Request for the ExportMandateTransactions method.

ExportMandateTransactionsResponse

Response for the ExportMandateTransactions method.

ExportMetadataTransactionsMetadata

Metadata for ExportMetadataTransactions.

ExportMetadataTransactionsRequest

Request for the ExportMetadataTransactions method.

ExportMetadataTransactionsResponse

Response for the ExportMetadataTransactions method.

FetchParticipantRequest

Request for the [FetchParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.FetchParticipant] method.

FinancialTransaction

A financial API transaction processed by the issuer switch. In UPI, this maps to the Pay API.

IssuerParticipant

A customer of the bank who participates in transactions processed by the issuer switch.

IssuerParticipants

Response for the [ActivateParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.ActivateParticipant], [DeactivateParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.DeactivateParticipant] and [MobileNumberChanged][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.MobileNumberChanged] methods.

ListComplaintTransactionsRequest

Request for the ListComplaintTransactions method. Callers can request for transactions to be filtered by the given filter criteria and specified pagination parameters.

ListComplaintTransactionsResponse

Response for the ListComplaintTransactionsResponse method.

ListFinancialTransactionsRequest

Request for the ListFinancialTransactions method. Callers can request for transactions to be filtered by the given filter criteria and specified pagination parameters.

ListFinancialTransactionsResponse

Response for the ListFinancialTransactions method.

ListMandateTransactionsRequest

Request for the ListMandateTransactions method. Callers can request for transactions to be filtered by the given filter criteria and specified pagination parameters.

ListMandateTransactionsResponse

Response for the ListMandateTransactionsResponse method.

ListMetadataTransactionsRequest

Request for the ListMetadataTransactions method. Callers can request for transactions to be filtered by the given filter criteria and specified pagination parameters.

ListMetadataTransactionsResponse

Response for the ListMetadataTransactions method.

ListRuleMetadataRequest

Request body for the ListRuleMetadata method.

ListRuleMetadataResponse

Response body for the ListRuleMetadata method.

ListRuleMetadataValuesRequest

Request body for the ListRuleMetadataValues method.

ListRuleMetadataValuesResponse

Response body for ListRuleMetadataValues. Contains a List of values for a given rule metadata resource.

ListRulesRequest

Request body for the ListRules method.

ListRulesResponse

Response body for the ListRules method.

MandateTransaction

A mandate processed by the issuer switch. In UPI, this maps to the Mandate API.

MerchantAdditionalInfo

Additional merchant information specific to India’s UPI requirements.

MerchantInfo

A merchant entity participating in a payment settlement transaction.

MerchantName

The name of a merchant who is a participant in a payment settlement transaction. Includes multiple possible names for the merchant.

MetadataTransaction

A metadata API transaction processed by the issuer switch. This includes UPI APIs such as List Accounts, Balance Enquiry, etc.

OriginalTransaction

Details of original transaction.

Participant

A participant in a transaction processed by the issuer switch.

ParticipantStateChangeRequest

Request for the [ActivateParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.ActivateParticipant], [DeactivateParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.DeactivateParticipant] and [MobileNumberUpdated][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.MobileNumberChanged] methods.

RaiseComplaintAdjustment

The adjusment flag and reason code for raising complaint.

RaiseDisputeAdjustment

The adjusment flag and reason code for raising dispute.

ResolveComplaintAdjustment

The adjusment flag and reason code for resolving the complaint.

ResolveComplaintMetadata

Metadata for ResolveComplaint.

ResolveComplaintRequest

Request for the ResolveComplaint method.

ResolveDisputeAdjustment

The adjusment flag and reason code for resolving the dispute.

ResolveDisputeMetadata

Metadata for ResolveDispute.

ResolveDisputeRequest

Request for the ResolveDispute method.

Rule

A rule that is executed by the issuer switch while processing an API transaction.

RuleMetadata

The metadata associated with a rule. This defines data that are used by the rule during execution.

RuleMetadataValue

Represent a single value in a rule’s metadata.

SettlementParticipant

A participant in a payment settlement transaction processed by the issuer switch. The participant could either be the payer or the payee in the transaction.

TransactionInfo

Information about a transaction processed by the issuer switch. The fields in this type are common across both financial and metadata transactions.

UpdateIssuerParticipantRequest

Request for the [UpdateIssuerParticipant][google.cloud.paymentgateway.issuerswitch.v1.IssuerSwitchParticipants.UpdateIssuerParticipant] method.

UpiTransaction

The payload for the log entry.

Enums

ApiType

The API type for a transaction. Every transaction processed by the issuer switch will be one of these API types.

TransactionSubType

The subtype of the complaint or dispute.

TransactionType

The type of a transaction. Every transaction processed by the issuer switch will be one of these transaction types. Transaction types are associated with a particular API type. This associated is documented with each value.

XmlApiType

XmlApiType specifies the API type of the request or response as specified in the XML payload.