Enum google_api_proto::google::cloud::paymentgateway::issuerswitch::v1::XmlApiType

#[repr(i32)]
pub enum XmlApiType {

    Unspecified = 0,
    ReqBalEnq = 1,
    ReqChkTxn = 2,
    ReqComplaint = 3,
    ReqHbt = 4,
    ReqListAccount = 5,
    ReqMandate = 6,
    ReqMandateConfirmation = 7,
    ReqOtp = 8,
    ReqPay = 9,
    ReqRegMob = 10,
    ReqSetCre = 11,
    ReqValCust = 12,
    ReqVoucher = 13,
    ReqVoucherConfirmation = 14,
    ReqTxnConfirmation = 15,
    RespBalEnq = 16,
    RespChkTxn = 17,
    RespComplaint = 18,
    RespHbt = 19,
    RespListAccount = 20,
    RespMandate = 21,
    RespMandateConfirmation = 22,
    RespOtp = 23,
    RespPay = 24,
    RespRegMob = 25,
    RespSetCre = 26,
    RespValCust = 27,
    RespVoucher = 28,
    RespVoucherConfirmation = 29,
    RespTxnConfirmation = 30,
    ReqActivation = 31,
    RespActivation = 32,

}

XmlApiType specifies the API type of the request or response as specified in the XML payload.

Variants

Unspecified = 0

Unspecified API type.

ReqBalEnq = 1

Balance enquiry request API type. Maps to UPI’s ReqBalEnq API.

ReqChkTxn = 2

Check transaction request API type. Maps to UPI’s ReqChkTxn API.

ReqComplaint = 3

Complaint request API type. Maps to UPI’s ReqComplaint API.

ReqHbt = 4

Heart beat request API type. Maps to UPI’s ReqHbt API.

ReqListAccount = 5

List accounts request API type. Maps to UPI’s ReqListAccount API.

ReqMandate = 6

Mandate request API. Maps to UPI’s ReqMandate API.

ReqMandateConfirmation = 7

Mandate confirmation request API type. Maps to UPI’s ReqMandateConfirmation API.

ReqOtp = 8

OTP request API. Maps to UPI’s ReqOtp API.

ReqPay = 9

Payment settlement request API type. Maps to UPI’s ReqPay API.

ReqRegMob = 10

Register mobile request API type. Maps to UPI’s ReqRegMob API.

ReqSetCre = 11

Update credentials request API type. Maps to UPI’s ReqSetCre API.

ReqValCust = 12

Validate customer request API type. Maps to UPI’s ReqValCust.

ReqVoucher = 13

Create voucher request API type. Maps to UPI’s ReqVoucher.

ReqVoucherConfirmation = 14

Voucher confirmation request API type. Maps to UPI’s ReqVoucherConfirmation API.

ReqTxnConfirmation = 15

Transaction confirmation request API type. Maps to UPI’s ReqTxnConfirmation API.

RespBalEnq = 16

Balance enquiry response API type. Maps to UPI’s RespBalEnq API.

RespChkTxn = 17

Check transaction response API type. Maps to UPI’s RespChkTxn API.

RespComplaint = 18

Complaint response API type. Maps to UPI’s RespComplaint API.

RespHbt = 19

Heart beat response API type. Maps to UPI’s RespHbt API.

RespListAccount = 20

List accounts response API type. Maps to UPI’s RespListAccount API.

RespMandate = 21

Mandate response API type. Maps to UPI’s RespMandate API.

RespMandateConfirmation = 22

Mandate confirmation response API type. Maps to UPI’s RespMandateConfirmation API.

RespOtp = 23

OTP response API. Maps to UPI’s RespOtp API.

RespPay = 24

Payment settlement response API type. Maps to UPI’s RespPay API.

RespRegMob = 25

Register mobile response API type. Maps to UPI’s RespRegMob API.

RespSetCre = 26

Update credentials response API type. Maps to UPI’s RespSetCre API.

RespValCust = 27

Validate customer response API type. Maps to UPI’s RespValCust.

RespVoucher = 28

Create voucher response API type. Maps to UPI’s RespVoucher.

RespVoucherConfirmation = 29

Voucher confirmation responseAPI type. Maps to UPI’s RespVoucherConfirmation API.

RespTxnConfirmation = 30

Transaction confirmation response API type. Maps to UPI’s RespTxnConfirmation API.

ReqActivation = 31

Activation request API type. Maps to UPI’s ReqActivation API.

RespActivation = 32

Activation response API type. Maps to UPI’s RespActivation API.

Implementations

impl XmlApiType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of XmlApiType.

pub fn from_i32(value: i32) -> Option<XmlApiType>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a XmlApiType, or None if value is not a valid variant.

impl XmlApiType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.