Struct google_api_proto::google::cloud::paymentgateway::issuerswitch::v1::issuer_switch_transactions_client::IssuerSwitchTransactionsClient

pub struct IssuerSwitchTransactionsClient<T> { /* private fields */ }

Fetch the issuer switch participant. Lists and exports transactions processed by the issuer switch.

Implementations

impl<T> IssuerSwitchTransactionsClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> IssuerSwitchTransactionsClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_metadata_transactions( &mut self, request: impl IntoRequest<ListMetadataTransactionsRequest>, ) -> Result<Response<ListMetadataTransactionsResponse>, Status>

List metadata transactions that satisfy the specified filter criteria.

pub async fn list_financial_transactions( &mut self, request: impl IntoRequest<ListFinancialTransactionsRequest>, ) -> Result<Response<ListFinancialTransactionsResponse>, Status>

List financial transactions that satisfy specified filter criteria.

pub async fn list_mandate_transactions( &mut self, request: impl IntoRequest<ListMandateTransactionsRequest>, ) -> Result<Response<ListMandateTransactionsResponse>, Status>

List mandate transactions that satisfy specified filter criteria.

pub async fn list_complaint_transactions( &mut self, request: impl IntoRequest<ListComplaintTransactionsRequest>, ) -> Result<Response<ListComplaintTransactionsResponse>, Status>

List complaint transactions that satisfy specified filter criteria.

pub async fn export_financial_transactions( &mut self, request: impl IntoRequest<ExportFinancialTransactionsRequest>, ) -> Result<Response<Operation>, Status>

Export financial transactions received within the specified time range as a file into a configured target location. The returned Operation type has the following method-specific fields:

metadata: [ExportFinancialTransactionsMetadata][google.cloud.paymentgateway.issuerswitch.v1.ExportFinancialTransactionsMetadata]

response: [ExportFinancialTransactionsResponse][google.cloud.paymentgateway.issuerswitch.v1.ExportFinancialTransactionsResponse]

The exported file will be in the standard CSV format where each row in the file represents a transaction. The file has the following fields in order:

TransactionID

Min Length - 35 characters

Max Length - 35 characters

Description - UPI transaction ID.

TransactionType

Min Length - 22 characters

Max Length - 25 characters

Description - Type of the transaction. This will be one of TRANSACTION_TYPE_CREDIT, TRANSACTION_TYPE_DEBIT or TRANSACTION_TYPE_REVERSAL.

TransactionSubType

Min Length - 3 characters

Max Length - 7 characters

Description - Subtype of the transaction. This will be one of COLLECT, or PAY.

CreationTime

Min Length - 20 characters

Max Length - 20 characters

Description - Timestamp (in UTC) indicating when the issuer switch created the transaction resource for processing the transaction. The format will be as per RFC-3339. Example : 2022-11-22T23:00:05Z

State

Min Length - 6 characters

Max Length - 9 characters

Description - State of the transaction. This will be one of FAILED, SUCCEEDED, or TIMED_OUT.

RRN

Min Length - 12 characters

Max Length - 12 characters

Description - Retrieval reference number associated with the transaction.

PayerVPA

Min Length - 3 characters

Max Length - 255 characters

Description - Virtual Payment Address (VPA) of the payer.

PayerMobileNumber

Min Length - 12 characters

Max Length - 12 characters

Description - Mobile number of the payer.

PayerIFSC

Min Length - 11 characters

Max Length - 11 characters

Description - IFSC of the payer’s bank account.

PayerAccountNumber

Min Length - 1 characters

Max Length - 30 characters

Description - Payer’s bank account number.

PayerAccountType

Min Length - 3 characters

Max Length - 7 characters

Description - Payer’s bank account type. This will be one of SAVINGS, DEFAULT, CURRENT, NRE, NRO, PPIWALLET, BANKWALLET, CREDIT, SOD, or UOD.

PayeeVPA

Min Length - 3 characters

Max Length - 255 characters

Description - Virtual Payment Address (VPA) of the payee.

PayeeMobileNumber

Min Length - 12 characters

Max Length - 12 characters

Description - Payee’s mobile number.

PayeeIFSC

Min Length - 11 characters

Max Length - 11 characters

Description - IFSC of the payee’s bank account.

PayeeAccountNumber

Min Length - 1 characters

Max Length - 30 characters

Description - Payee’s bank account number.

PayeeAccountType

Min Length - 3 characters

Max Length - 10 characters

Description - Payee’s bank account type. This will be one of SAVINGS, DEFAULT, CURRENT, NRE, NRO, PPIWALLET, BANKWALLET, CREDIT, SOD, or UOD.

PayeeMerchantID

Min Length - 1 characters

Max Length - 255 characters

Description - Payee’s merchant ID, only if the payee is a merchant.

PayeeMerchantName

Min Length - 1 characters

Max Length - 99 characters

Description - Payee’s merchant name, only if the payee is a merchant.

PayeeMCC

Min Length - 4 characters

Max Length - 4 characters

Description - Payee’s Merchant Category Code (MCC), only if the payee is a merchant.

Currency

Min Length - 3 characters

Max Length - 3 characters

Description - Currency of the amount involved in the transaction. The currency codes are defined in ISO 4217.

Amount

Description - Amount involved in the transaction.

AdapterRequestIDs

Min Length - 0 characters

Max Length - 2,000 characters

Description - List of adapter request IDs (colon separated) used when invoking the Adapter APIs for fulfilling a transaction request.

ErrorCode

Min Length - 0 characters

Max Length - 255 characters

Description - Error code of a failed transaction.

ErrorMessage

Min Length - 0 characters

Max Length - 10,000 characters

Description - Error description for a failed transaction.

UPIErrorCode

Min Length - 0 characters

Max Length - 3 characters

Description - Error code as per the UPI specification. The issuer switch maps the ErrorCode to an appropriate error code that complies with the UPI specification.

PayerDeviceInfoTypeAppName

Min Length - 0 characters

Max Length - 20 characters

Description - Payment application name on the payer’s device.

PayerDeviceInfoTypeCapability

Min Length - 0 characters

Max Length - 99 characters

Description - Capability of the payer’s device.

PayerDeviceInfoTypeGeoCode

Min Length - 0 characters

Max Length - 15 characters

Description - Geo code of the payer’s device. This will include floating point values for latitude and longitude (separated by colon).

PayerDeviceInfoTypeID

Min Length - 0 characters

Max Length - 35 characters

Description - Device ID of the payer’s device.

PayerDeviceInfoTypeIP

Min Length - 0 characters

Max Length - 39 characters

Description - IP address of the payer’s device.

PayerDeviceInfoTypeLocation

Min Length - 0 characters

Max Length - 40 characters

Description - Coarse location of the payer’s device.

PayerDeviceInfoTypeOS

Min Length - 0 characters

Max Length - 20 characters

Description - Operating system on the payer’s device.

PayerDeviceInfoTypeTelecomProvider

Min Length - 0 characters

Max Length - 99 characters

Description - Telecom provider for the payer’s device.

PayerDeviceInfoTypeDeviceType

Min Length - 0 characters

Max Length - 9 characters

Description - Type of the payer’s device. This will be one of ‘MOB’, ‘INET’, ‘USDC/USDB’, ‘POS’.

PayeeDeviceInfoTypeAppName

Min Length - 0 characters

Max Length - 20 characters

Description - Payment application name on the payee’s device.

PayeeDeviceInfoTypeCapability

Min Length - 0 characters

Max Length - 99 characters

Description - Capability of the payee’s device.

PayeeDeviceInfoTypeGeoCode

Min Length - 0 characters

Max Length - 15 characters

Description - Geo code of the payee’s device. This will include floating point values for latitude and longitude (separated by colon).

PayeeDeviceInfoTypeID

Min Length - 0 characters

Max Length - 35 characters

Description - Device ID of the payee’s device.

PayeeDeviceInfoTypeIP

Min Length - 0 characters

Max Length - 39 characters

Description - IP address of the payee’s device.

PayeeDeviceInfoTypeLocation

Min Length - 0 characters

Max Length - 40 characters

Description - Coarse location of the payee’s device.

PayeeDeviceInfoTypeOS

Min Length - 0 characters

Max Length - 20 characters

Description - Operating system on the payee’s device.

PayeeDeviceInfoTypeTelecomProvider

Min Length - 0 characters

Max Length - 99 characters

Description - Telecom provider for the payee’s device.

PayeeDeviceInfoTypeDeviceType

Min Length - 0 characters

Max Length - 9 characters

Description - Type of the payee’s device. This will be one of ‘MOB’, ‘INET’, ‘USDC/USDB’, ‘POS’.

ReferenceID

Min Length - 0 characters

Max Length - 35 characters

Description - Consumer reference number to identify loan number, order id etc.

ReferenceURI

Min Length - 1 characters

Max Length - 35 characters

Description - URL for the transaction.

ReferenceCategory

Min Length - 2 characters

Max Length - 2 characters

Description - Reference category.

pub async fn export_metadata_transactions( &mut self, request: impl IntoRequest<ExportMetadataTransactionsRequest>, ) -> Result<Response<Operation>, Status>

Export metadata transactions received within the specified time range as a file into a configured target location. The returned Operation type has the following method-specific fields:

metadata: [ExportMetadataTransactionsMetadata][google.cloud.paymentgateway.issuerswitch.v1.ExportMetadataTransactionsMetadata]

response: [ExportMetadataTransactionsResponse][google.cloud.paymentgateway.issuerswitch.v1.ExportMetadataTransactionsResponse]

The exported file will be in the standard CSV format where each row in the file represents a transaction. The file has the following fields in order:

TransactionID

Min Length - 35 characters

Max Length - 35 characters

Description - UPI transaction ID.

APIType

Description - The transaction’s API type. The value will be of the [ApiType][google.cloud.paymentgateway.issuerswitch.v1.ApiType] enum.

TransactionType

Description - Type of the transaction. The value will be of the [TransactionType][google.cloud.paymentgateway.issuerswitch.v1.TransactionType] enum.

CreationTime

Min Length - 20 characters

Max Length - 20 characters

Description - Timestamp (in UTC) indicating when the issuer switch created the transaction resource for processing the transaction. The format will be as per RFC-3339. Example : 2022-11-22T23:00:05Z

State

Min Length - 6 characters

Max Length - 9 characters

Description - State of the transaction. This will be one of FAILED, SUCCEEDED, or TIMED_OUT.

OriginVPA

Min Length - 3 characters

Max Length - 255 characters

Description - Virtual Payment Address (VPA) of the originator of the transaction.

AdapterRequestIDs

Min Length - 0 characters

Max Length - 2,000 characters

Description - List of adapter request IDs (colon separated) used when invoking the Adapter APIs for fulfilling a transaction request.

ErrorCode

Min Length - 0 characters

Max Length - 255 characters

Description - Error code of the failed transaction.

ErrorMessage

Min Length - 0 characters

Max Length - 10,000 characters

Description - Error description for the failed transaction.

UPIErrorCode

Min Length - 0 characters

Max Length - 3 characters

Description - Error code as per the UPI specification. The issuer switch maps the ErrorCode to an appropriate error code that complies with the UPI specification.

pub async fn export_mandate_transactions( &mut self, request: impl IntoRequest<ExportMandateTransactionsRequest>, ) -> Result<Response<Operation>, Status>

Export mandate transactions received within the specified time range as a file into a configured target location. The returned Operation type has the following method-specific fields:

metadata: [ExportMandateTransactionsMetadata][google.cloud.paymentgateway.issuerswitch.v1.ExportMandateTransactionsMetadata]

response: [ExportMandateTransactionsResponse][google.cloud.paymentgateway.issuerswitch.v1.ExportMandateTransactionsResponse]

The exported file will be in the standard CSV format where each row in the file represents a transaction. The file has the following fields in order:

TransactionID

Min Length - 35 characters

Max Length - 35 characters

Description - UPI transaction ID.

UniqueMandateNumber

Min Length - 3 characters

Max Length - 70 characters

Description - UPI Unique Mandate Number.

TransactionType

Min Length - 23 characters

Max Length - 23 characters

Description - Type of the transaction. This will be one of TRANSACTION_TYPE_CREATE, TRANSACTION_TYPE_REVOKE, TRANSACTION_TYPE_UPDATE, TRANSACTION_TYPE_PAUSE or TRANSACTION_TYPE_UNPAUSE.

CreationTime

Min Length - 20 characters

Max Length - 20 characters

Description - Timestamp (in UTC) indicating when the issuer switch created the transaction resource for processing the transaction. The format will be as per RFC-3339. Example : 2022-11-22T23:00:05Z

State

Min Length - 6 characters

Max Length - 9 characters

Description - State of the transaction. This will be one of FAILED, SUCCEEDED, or TIMED_OUT.

PayerVPA

Min Length - 3 characters

Max Length - 255 characters

Description - Virtual Payment Address (VPA) of the payer.

PayerMobileNumber

Min Length - 12 characters

Max Length - 12 characters

Description - Mobile number of the payer.

PayerIFSC

Min Length - 11 characters

Max Length - 11 characters

Description - IFSC of the payer’s bank account.

PayerAccountNumber

Min Length - 1 characters

Max Length - 30 characters

Description - Payer’s bank account number.

PayerAccountType

Min Length - 3 characters

Max Length - 7 characters

Description - Payer’s bank account type. This will be one of SAVINGS, DEFAULT, CURRENT, NRE, NRO, PPIWALLET, BANKWALLET, CREDIT, SOD, or UOD.

PayeeVPA

Min Length - 3 characters

Max Length - 255 characters

Description - Virtual Payment Address (VPA) of the payee.

PayeeMobileNumber

Min Length - 12 characters

Max Length - 12 characters

Description - Mobile number of the payee.

PayeeIFSC

Min Length - 11 characters

Max Length - 11 characters

Description - IFSC of the payee’s bank account.

PayeeAccountNumber

Min Length - 1 characters

Max Length - 30 characters

Description - Payee’s bank account number.

PayeeAccountType

Min Length - 3 characters

Max Length - 10 characters

Description - Payee’s bank account type. This will be one of SAVINGS, DEFAULT, CURRENT, NRE, NRO, PPIWALLET, BANKWALLET, CREDIT, SOD, or UOD.

PayeeMerchantID

Min Length - 1 characters

Max Length - 30 characters

Description - Payee’s merchant ID, only if the payee is a merchant

PayeeMerchantName

Min Length - 1 characters

Max Length - 99 characters

Description - Payee’s merchant name, only if the payee is a merchant.

PayeeMCC

Min Length - 4 characters

Max Length - 4 characters

Description - Payee’s Merchant Category Code (MCC), only if the payee is a merchant.

Amount

Description - Amount specified in the mandate.

RecurrencePattern

Description - Reccurence pattern of the mandate. The value will be of the [MandateTransaction.RecurrencePatternType][google.cloud.paymentgateway.issuerswitch.v1.MandateTransaction.RecurrencePatternType] enum.

RecurrenceRuleType

Description - Reccurrence rule type of the mandate. The value will be of the [MandateTransaction.RecurrenceRuleType][google.cloud.paymentgateway.issuerswitch.v1.MandateTransaction.RecurrenceRuleType] enum.

RecurrenceRuleValue

Min Length - 0 characters

Max Length - 2 characters

Description - Recurrence rule value of the mandate. This will be an integer between 1 and 31.

Revokeable

Min Length - 4 characters

Max Length - 5 characters

Description - Boolean value specifying if the mandate is revokable.

StartDate

Min Length - 10 characters

Max Length - 10 characters

Description - The start date of the mandate in DD-MM-YYYY format.

EndDate

Min Length - 10 characters

Max Length - 10 characters

Description - The end date of the mandate in DD-MM-YYYY format.

AmountRuleType

Description - The amount rule of the mandate. The value will be of the [MandateTransaction.AmountRuleType][google.cloud.paymentgateway.issuerswitch.v1.MandateTransaction.AmountRuleType] enum.

ApprovalReference

Min Length - 6 characters

Max Length - 9 characters

Description - The block funds reference generated by the bank, if funds have been blocked for the mandate. This column will have a value only when the RecurrencePattern is ONETIME.

BlockFunds

Min Length - 4 characters

Max Length - 5 characters

Description - Boolean value specifying if the mandate transaction requested to block funds.

LastUpdateTime

Min Length - 20 characters

Max Length - 20 characters

Description - Timestamp (in UTC) indicating when was the last modification made to the mandate. The format will be as per RFC-3339. Example : 2022-11-22T23:00:05Z

AdapterRequestIDs

Min Length - 0 characters

Max Length - 2,000 characters

Description - List of adapter request IDs (colon separated) used when invoking the Adapter APIs for fulfilling a transaction request.

ErrorCode

Min Length - 0 characters

Max Length - 255 characters

Description - Error code of the failed transaction.

ErrorMessage

Min Length - 0 characters

Max Length - 10,000 characters

Description - Error description for the failed transaction.

UPIErrorCode

Min Length - 0 characters

Max Length - 3 characters

Description - Error code as per the UPI specification. The issuer switch maps the ErrorCode to an appropriate error code that complies with the UPI specification.

PayerDeviceInfoTypeAppName

Min Length - 0 characters

Max Length - 20 characters

Description - Payment application name on the payer’s device.

PayerDeviceInfoTypeCapability

Min Length - 0 characters

Max Length - 99 characters

Description - Capability of the payer’s device.

PayerDeviceInfoTypeGeoCode

Min Length - 0 characters

Max Length - 15 characters

Description - Geo code of the payer’s device. This will include floating point values for latitude and longitude (separated by colon).

PayerDeviceInfoTypeID

Min Length - 0 characters

Max Length - 35 characters

Description - Device ID of the payer’s device.

PayerDeviceInfoTypeIP

Min Length - 0 characters

Max Length - 39 characters

Description - IP address of the payer’s device.

PayerDeviceInfoTypeLocation

Min Length - 0 characters

Max Length - 40 characters

Description - Coarse location of the payer’s device.

PayerDeviceInfoTypeOS

Min Length - 0 characters

Max Length - 20 characters

Description - Operating system on the payer’s device.

PayerDeviceInfoTypeTelecomProvider

Min Length - 0 characters

Max Length - 99 characters

Description - Telecom provider for the payer’s device.

PayerDeviceInfoTypeDeviceType

Min Length - 0 characters

Max Length - 9 characters

Description - Type of the payer’s device. This will be one of ‘MOB’, ‘INET’, ‘USDC/USDB’, ‘POS’.

PayeeDeviceInfoTypeAppName

Min Length - 0 characters

Max Length - 20 characters

Description - Payment application name on the payee’s device.

PayeeDeviceInfoTypeCapability

Min Length - 0 characters

Max Length - 99 characters

Description - Capability of the payee’s device.

PayeeDeviceInfoTypeGeoCode

Min Length - 0 characters

Max Length - 15 characters

Description - Geo code of the payee’s device. This will include floating point values for latitude and longitude (separated by colon).

PayeeDeviceInfoTypeID

Min Length - 0 characters

Max Length - 35 characters

Description - Device ID of the payee’s device.

PayeeDeviceInfoTypeIP

Min Length - 0 characters

Max Length - 39 characters

Description - IP address of the payee’s device.

PayeeDeviceInfoTypeLocation

Min Length - 0 characters

Max Length - 40 characters

Description - Coarse location of the payee’s device.

PayeeDeviceInfoTypeOS

Min Length - 0 characters

Max Length - 20 characters

Description - Operating system on the payee’s device.

PayeeDeviceInfoTypeTelecomProvider

Min Length - 0 characters

Max Length - 99 characters

Description - Telecom provider for the payee’s device.

PayeeDeviceInfoTypeDeviceType

Min Length - 0 characters

Max Length - 9 characters

Description - Type of the payee’s device. This will be one of MOB, INET, USDC/USDB, POS.

ReferenceID

Min Length - 0 characters

Max Length - 35 characters

Description - Consumer reference number to identify loan number, order id etc.

ReferenceURI

Min Length - 1 characters

Max Length - 35 characters

Description - URL for the transaction.

ReferenceCategory

Min Length - 2 characters

Max Length - 2 characters

Description - Reference category.

MandateName

Min Length - 1 characters

Max Length - 255 characters

Description - The mandate’s name.

pub async fn export_complaint_transactions( &mut self, request: impl IntoRequest<ExportComplaintTransactionsRequest>, ) -> Result<Response<Operation>, Status>

Export complaint transactions received within the specified time range as a file into a configured target location. The returned Operation type has the following method-specific fields:

metadata: [ExportComplaintTransactionsMetadata][google.cloud.paymentgateway.issuerswitch.v1.ExportComplaintTransactionsMetadata]

response: [ExportComplaintTransactionsResponse][google.cloud.paymentgateway.issuerswitch.v1.ExportComplaintTransactionsResponse]

The exported file will be in the standard CSV format where each row in the file represents a transaction. The file has the following fields in order:

TransactionID

Min Length - 35 characters

Max Length - 35 characters

Description - UPI transaction ID.

TransactionType

Min Length - 23 characters

Max Length - 30 characters

Description - Type of the transaction. This will be one of TRANSACTION_TYPE_CHECK_STATUS, TRANSACTION_TYPE_COMPLAINT, TRANSACTION_TYPE_REVERSAL, TRANSACTION_TYPE_DISPUTE, TRANSACTION_TYPE_REFUND, or TRANSACTION_TYPE_STATUS_UPDATE.

CreationTime

Min Length - 20 characters

Max Length - 20 characters

Description - Timestamp (in UTC) indicating when the issuer switch created the transaction resource for processing the transaction. The format will be as per RFC-3339. Example : 2022-11-22T23:00:05Z 1: State

Min Length - 6 characters

Max Length - 9 characters

Description - State of the transaction. This will be one of FAILED, SUCCEEDED, or TIMED_OUT.

OriginalRRN

Min Length - 12 characters

Max Length - 12 characters

Description - Retrieval reference number of the original payment transaction.

BankType

Min Length - 8 characters

Max Length - 11 characters

Description - The subtype of the transaction based on the bank involved. This will be one of BENEFICIARY, or REMITTER.

OriginalTransactionID

Min Length - 35 characters

Max Length - 35 characters

Description - Transaction ID of the original unresolved transaction.

RaiseComplaintAdjFlag

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the type of action to raise the complaint.

RaiseComplaintAdjCode

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the reason of action to raise the complaint.

ResolveComplaintAdjFlag

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the type of action to resolve the complaint.

ResolveComplaintAdjCode

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the reason of action to resolve the complaint.

RaiseDisputeAdjFlag

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the type of action to raise the dispute.

RaiseDisputeAdjCode

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the reason of action to raise the dispute.

ResolveDisputeAdjFlag

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the type of action to resolve the dispute.

ResolveDisputeAdjCode

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the reason of action to resolve the dispute.

Amount

Description - Amount to be resolved.

CurrentCycle

Min Length - 4 characters

Max Length - 5 characters

Description - Boolean value specifying if the complaint / dispute belongs to current settlement cycle or not.

CRN

Min Length - 0 characters

Max Length - 255 characters

Description - Defines the Complaint Reference number.

AdjTime

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the time when the resolution was done.

RespAdjFlag

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the response category type.

RespAdjCode

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the response reason used.

AdjRemarks

Min Length - 0 characters

Max Length - 255 characters

Description - Indicates the additional remarks for the complaint / dispute.

AdapterRequestIDs

Min Length - 0 characters

Max Length - 2,000 characters

Description - List of adapter request IDs (colon separated) used when invoking the Adapter APIs for fulfilling a transaction request.

ErrorCode

Min Length - 0 characters

Max Length - 255 characters

Description - Error code of the failed transaction.

ErrorMessage

Min Length - 0 characters

Max Length - 10,000 characters

Description - Error description for the failed transaction.

UPIErrorCode

Min Length - 0 characters

Max Length - 3 characters

Description - Error code as per the UPI specification. The issuer switch service maps the ErrorCode to an appropriate error code that complies with the UPI specification.