Enum google_api_proto::google::cloud::paymentgateway::issuerswitch::v1::resolve_dispute_adjustment::ReasonCode

#[repr(i32)]
pub enum ReasonCode {

    Unspecified = 0,
    ChargebackBeneficiaryCannotCreditOrPreArbitrationDuplicateProcess = 1,
    PreArbitrationDeclinedBeneficiaryCreditedOnline = 3,
    PreArbitrationDeclinedBeneficiaryCreditedManually = 4,
    DeferredChargebackAcceptanceAccountNotCreditedTccRaised = 5,
    DeferredRePresentmentRaiseAccountCreditedTccRaised = 6,
    DeferredPreArbitrationAcceptanceAccountNotCredited = 7,
    DeferredPreArbitrationDeclinedAccountCredited = 8,
    FraudChargebackAcceptAmountRecoveredFromFraudulentAccount = 9,
    FraudChargebackRepresentmentLienMarkedInsufficientBalance = 10,
    FraudChargebackRepresentmentFirNotProvided = 11,
    FraudChargebackRepresentmentReasonOthers = 12,
    RePresentmentRaiseBeneficiaryCreditedOnline = 13,
    RePresentmentRaiseBeneficiaryCreditedManually = 14,
    CreditAdjustmentGoodsServicesCreditNotProcessed = 15,
    CreditAdjustmentGoodsServicesDefective = 16,
    CreditAdjustmentPaidByAlternateMeans = 17,
    CreditAdjustmentGoodsServicesNotReceived = 18,
    CreditAdjustmentMerchantNotReceivedConfirmation = 19,
    CreditAdjustmentDuplicateTransaction = 20,
    CreditAdjustmentReasonOthers = 21,
    CreditAdjustmentNonMatchingAccountNumber = 22,
    CreditAdjustmentCardHolderChargedMore = 23,
    CreditAdjustmentCreditNotProcessed = 24,
    CreditAdjustmentBeneficiaryCannotCredit = 25,
    ChargebackAcceptanceMerchantCannotProvideService = 26,
    RePresentmentRaiseGoodsServicesProvided = 27,
    PreArbitrationDeclinedServicesProvidedLater = 28,
    PreArbitrationAcceptanceServicesNotProvidedByMerchant = 29,
    ArbitrationAcceptanceIllegibleFulfilment = 30,
    ArbitrationContinuationCustomerStillNotReceivedService = 31,
    ArbitrationWithdrawnCustomerReceivedServiceLater = 32,
    ArbitrationVerdictPanelVerdict = 33,
    ManualAdjustmentReason = 34,
    AttributingCustomer = 35,
    AttributingTechnicalIssue = 36,
    WrongCreditChargebackAcceptanceAmountRecovered = 37,
    WrongCreditRepresentmentLienMarkedInsufficientBalance = 38,
    WrongCreditRepresentmentCustomerInaccessible = 39,
    WrongCreditRepresentmentReasonOthers = 40,

}

The dispute resolution reason code.

Variants

Unspecified = 0

Unspecified reason code.

ChargebackBeneficiaryCannotCreditOrPreArbitrationDuplicateProcess = 1

Beneficiary bank unable to credit their customer account for Chargeback Acceptance dispute or duplicate processing for Pre Arbitration Acceptance dispute. This reason code maps to the 111 reason code as defined in NPCI’s UDIR specification.

PreArbitrationDeclinedBeneficiaryCreditedOnline = 3

Beneficiary account has been credited online. This reason code maps to the 112 reason code for Pre-arbitration declined dispute as defined in NPCI’s UDIR specification.

PreArbitrationDeclinedBeneficiaryCreditedManually = 4

Beneficiary account has been credited manually post reconciliation. This reason code maps to the 113 reason code for Pre-arbitration declined dispute as defined in NPCI’s UDIR specification.

DeferredChargebackAcceptanceAccountNotCreditedTccRaised = 5

Customer account is not credited, TCC raised inadvertently. This reason code maps to the 122 reason code as defined in NPCI’s UDIR specification.

DeferredRePresentmentRaiseAccountCreditedTccRaised = 6

Customer account is credited successfully and TCC raised accordingly. This reason code maps to the 123 reason code as defined in NPCI’s UDIR specification.

DeferredPreArbitrationAcceptanceAccountNotCredited = 7

Customer account is not credited, TCC and Re-Presentment raised inadvertently. This reason code maps to the 125 reason code as defined in NPCI’s UDIR specification.

DeferredPreArbitrationDeclinedAccountCredited = 8

Customer account is credited successfully and TCC and Re-Presentment raised accordingly. This reason code maps to the 126 reason code as defined in NPCI’s UDIR specification.

FraudChargebackAcceptAmountRecoveredFromFraudulentAccount = 9

Amount has been recovered successfully from the fraudulent customer account. This reason code maps to the 129 reason code as defined in NPCI’s UDIR specification.

FraudChargebackRepresentmentLienMarkedInsufficientBalance = 10

Lien marked however, customer account is not having sufficient balance to debit. This reason code maps to the 130 reason code for Fraud chargeback representment dispute as defined in NPCI’s UDIR specification.

FraudChargebackRepresentmentFirNotProvided = 11

FIR Copy not provided for the disputed transaction. This reason code maps to the 131 reason code as defined in NPCI’s UDIR specification.

FraudChargebackRepresentmentReasonOthers = 12

Other reason for Fraud chargeback representment dispute. This reason code maps to the 132 reason code as defined in NPCI’s UDIR specification.

RePresentmentRaiseBeneficiaryCreditedOnline = 13

Beneficiary account credited online. This reason code maps to the 208 reason code for Re-presentment raise dispute as defined in NPCI’s UDIR specification.

RePresentmentRaiseBeneficiaryCreditedManually = 14

Beneficiary account credited manually post reconciliation. This reason code maps to the 209 reason code for Re-presentment raise dispute as defined in NPCI’s UDIR specification.

CreditAdjustmentGoodsServicesCreditNotProcessed = 15

Credit not processed for cancelled or returned goods and services. This reason code maps to the 1061 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentGoodsServicesDefective = 16

Goods and Services not as described / defective. This reason code maps to the 1062 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentPaidByAlternateMeans = 17

Paid by alternate means. This reason code maps to the 1063 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentGoodsServicesNotReceived = 18

Goods or Services Not Provided / Not Received. This reason code maps to the 1064 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentMerchantNotReceivedConfirmation = 19

Account debited but transaction confirmation not received at merchant location. This reason code maps to the 1065 reason code for Credit adjustment as defined in NPCI’s UDIR specification.

CreditAdjustmentDuplicateTransaction = 20

Duplicate /Multiple Transaction. This reason code maps to the 1084 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentReasonOthers = 21

Other reason for Credit adjustment. This reason code maps to the 1090 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentNonMatchingAccountNumber = 22

Non Matching account number. This reason code maps to the 1091 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentCardHolderChargedMore = 23

Card holder was charged more than the transaction amount. This reason code maps to the 1092 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentCreditNotProcessed = 24

Credit not Processed. This reason code maps to the 1093 reason code as defined in NPCI’s UDIR specification.

CreditAdjustmentBeneficiaryCannotCredit = 25

Beneficiary bank unable to credit their customer account. This reason code maps to the 1094 reason code for Credit Adjustment dispute as defined in NPCI’s UDIR specification.

ChargebackAcceptanceMerchantCannotProvideService = 26

Merchant was unable to provide the service. This reason code maps to the 1095 reason code as defined in NPCI’s UDIR specification.

RePresentmentRaiseGoodsServicesProvided = 27

Services/Goods provided see the supporting document. This reason code maps to the 1096 reason code as defined in NPCI’s UDIR specification.

PreArbitrationDeclinedServicesProvidedLater = 28

Services provided later see supporting documents. This reason code maps to the 1098 reason code as defined in NPCI’s UDIR specification.

PreArbitrationAcceptanceServicesNotProvidedByMerchant = 29

Services not provided by the merchant. This reason code maps to the 1099 reason code as defined in NPCI’s UDIR specification.

ArbitrationAcceptanceIllegibleFulfilment = 30

Illegible Fulfilment. This reason code maps to the 1101 reason code for arbitration acceptance dispute as defined in NPCI’s UDIR specification.

ArbitrationContinuationCustomerStillNotReceivedService = 31

Customer has still not received the service. This reason code maps to the 1102 reason code as defined in NPCI’s UDIR specification.

ArbitrationWithdrawnCustomerReceivedServiceLater = 32

Customer has received the service later. This reason code maps to the 1103 reason code as defined in NPCI’s UDIR specification.

ArbitrationVerdictPanelVerdict = 33

Panel will give the verdict. This reason code maps to the 1104 reason code as defined in NPCI’s UDIR specification.

ManualAdjustmentReason = 34

Manual adjustment. This reason code maps to the 2001 reason code as defined in NPCI’s UDIR specification.

AttributingCustomer = 35

Attributing to the Customer. This reason code maps to the AC reason code as defined in NPCI’s UDIR specification.

AttributingTechnicalIssue = 36

Attributing to the Technical issue at bank/aggregator/merchant. This reason code maps to the AT reason code as defined in NPCI’s UDIR specification.

WrongCreditChargebackAcceptanceAmountRecovered = 37

Amount has been recovered successfully from the unintended customer account. This reason code maps to the WC2 reason code as defined in NPCI’s UDIR specification.

WrongCreditRepresentmentLienMarkedInsufficientBalance = 38

Lien marked however customer account is not having sufficient balance to debit the customer account. This reason code maps to the WC3 reason code for Wrong credit representment dispute as defined in NPCI’s UDIR specification.

WrongCreditRepresentmentCustomerInaccessible = 39

Customer is not accessible for obtaining debit confirmation. This reason code maps to the WC4 reason code as defined in NPCI’s UDIR specification.

WrongCreditRepresentmentReasonOthers = 40

Other reason for Wrong credit representment. This reason code maps to the WC5 reason code as defined in NPCI’s UDIR specification.

Implementations

impl ReasonCode

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ReasonCode.

pub fn from_i32(value: i32) -> Option<ReasonCode>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a ReasonCode, or None if value is not a valid variant.

impl ReasonCode

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.