Module google_api_proto::google::cloud::paymentgateway::issuerswitch::v1::transaction_info

Nested message and enum types in TransactionInfo.

Modules

adapter_info

Nested message and enum types in AdapterInfo.

Structs

AdapterInfo

Information about an adapter invocation triggered as part of the processing of an API transaction.

TransactionErrorDetails

All details about any error in the processing of an API transaction.

TransactionMetadata

Common metadata about an API transaction.

TransactionRiskInfo

Information about the transaction’s risk evaluation as provided by the payments orchestrator.

Enums

State

Specifies the current state of the transaction.

TransactionSubType

The subtype of a transaction. This value is used only for certain API type and transaction type combinations.