Module google_api_proto::google::cloud::policytroubleshooter::iam::v3

Modules

allow_binding_explanation

Nested message and enum types in AllowBindingExplanation.

condition_context

Nested message and enum types in ConditionContext.

condition_explanation

Nested message and enum types in ConditionExplanation.

deny_rule_explanation

Nested message and enum types in DenyRuleExplanation.

policy_troubleshooter_client

Generated client implementations.

troubleshoot_iam_policy_response

Nested message and enum types in TroubleshootIamPolicyResponse.

Structs

AccessTuple

Information about the principal, resource, and permission to check.

AllowBindingExplanation

Details about how a role binding in an allow policy affects a principal’s ability to use a permission.

AllowPolicyExplanation

Details about how the relevant IAM allow policies affect the final access state.

ConditionContext

Additional context for troubleshooting conditional role bindings and deny rules.

ConditionExplanation

Explanation for how a condition affects a principal’s access

DenyPolicyExplanation

Details about how the relevant IAM deny policies affect the final access state.

DenyRuleExplanation

Details about how a deny rule in a deny policy affects a principal’s ability to use a permission.

ExplainedAllowPolicy

Details about how a specific IAM allow policy contributed to the final access state.

ExplainedDenyPolicy

Details about how a specific IAM deny policy [Policy][google.iam.v2.Policy] contributed to the access check.

ExplainedDenyResource

Details about how a specific resource contributed to the deny policy evaluation.

TroubleshootIamPolicyRequest

Request for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.iam.v3.PolicyTroubleshooter.TroubleshootIamPolicy].

TroubleshootIamPolicyResponse

Response for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.iam.v3.PolicyTroubleshooter.TroubleshootIamPolicy].

Enums

AllowAccessState

Whether IAM allow policies gives the principal the permission.

DenyAccessState

Whether IAM deny policies deny the principal the permission.

HeuristicRelevance

The extent to which a single data point contributes to an overall determination.

MembershipMatchingState

Whether the principal in the request matches the principal in the policy.

PermissionPatternMatchingState

Whether the permission in the request matches the permission in the policy.

RolePermissionInclusionState

Whether a role includes a specific permission.