Module google_api_proto::google::cloud::policytroubleshooter::v1

Modules

binding_explanation

Nested message and enum types in BindingExplanation.

iam_checker_client

Generated client implementations.

Structs

AccessTuple

Information about the principal, resource, and permission to check.

BindingExplanation

Details about how a binding in a policy affects a principal’s ability to use a permission.

ExplainedPolicy

Details about how a specific IAM [Policy][google.iam.v1.Policy] contributed to the access check.

TroubleshootIamPolicyRequest

Request for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.v1.IamChecker.TroubleshootIamPolicy].

TroubleshootIamPolicyResponse

Response for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.v1.IamChecker.TroubleshootIamPolicy].

Enums

AccessState

Whether a principal has a permission for a resource.

HeuristicRelevance

The extent to which a single data point, such as the existence of a binding or whether a binding includes a specific principal, contributes to an overall determination.