Module google_api_proto::google::cloud::privilegedaccessmanager::v1

Modules

approval_workflow

Nested message and enum types in ApprovalWorkflow.

check_onboarding_status_response

Nested message and enum types in CheckOnboardingStatusResponse.

entitlement

Nested message and enum types in Entitlement.

grant

Nested message and enum types in Grant.

justification

Nested message and enum types in Justification.

manual_approvals

Nested message and enum types in ManualApprovals.

privileged_access

Nested message and enum types in PrivilegedAccess.

privileged_access_manager_client

Generated client implementations.

search_entitlements_request

Nested message and enum types in SearchEntitlementsRequest.

search_grants_request

Nested message and enum types in SearchGrantsRequest.

Structs

AccessControlEntry

AccessControlEntry is used to control who can do some operation.

ApprovalWorkflow

Different types of approval workflows that can be used to gate privileged access granting.

ApproveGrantRequest

Request message for ApproveGrant method.

CheckOnboardingStatusRequest

Request message for CheckOnboardingStatus method.

CheckOnboardingStatusResponse

Response message for CheckOnboardingStatus method.

CreateEntitlementRequest

Message for creating an entitlement.

CreateGrantRequest

Message for creating a grant

DeleteEntitlementRequest

Message for deleting an entitlement.

DenyGrantRequest

Request message for DenyGrant method.

Entitlement

An entitlement defines the eligibility of a set of users to obtain predefined access for some time possibly after going through an approval workflow.

GetEntitlementRequest

Message for getting an entitlement.

GetGrantRequest

Message for getting a grant.

Grant

This is to ensure that the Grants and ProducerGrants proto are byte compatible. A grant represents a request from a user for obtaining the access specified in an entitlement they are eligible for.

Justification

Justification represents a justification for requesting access.

ListEntitlementsRequest

Message for requesting list of entitlements.

ListEntitlementsResponse

Message for response to listing entitlements.

ListGrantsRequest

Message for requesting list of grants.

ListGrantsResponse

Message for response to listing grants.

ManualApprovals

A manual approval workflow where users who are designated as approvers need to call the ApproveGrant/DenyGrant APIs for a grant. The workflow can consist of multiple serial steps where each step defines who can act as approver in that step and how many of those users should approve before the workflow moves to the next step.

OperationMetadata

Represents the metadata of the long-running operation.

PrivilegedAccess

Privileged access that this service can be used to gate.

RevokeGrantRequest

Request message for RevokeGrant method.

SearchEntitlementsRequest

Request message for SearchEntitlements method.

SearchEntitlementsResponse

Response message for SearchEntitlements method.

SearchGrantsRequest

Request message for SearchGrants method.

SearchGrantsResponse

Response message for SearchGrants method.

UpdateEntitlementRequest

Message for updating an entitlement.