Module google_api_proto::google::cloud::recaptchaenterprise::v1

Modules

account_defender_assessment

Nested message and enum types in AccountDefenderAssessment.

account_verification_info

Nested message and enum types in AccountVerificationInfo.

annotate_assessment_request

Nested message and enum types in AnnotateAssessmentRequest.

endpoint_verification_info

Nested message and enum types in EndpointVerificationInfo.

event

Nested message and enum types in Event.

firewall_action

Nested message and enum types in FirewallAction.

fraud_prevention_assessment

Nested message and enum types in FraudPreventionAssessment.

fraud_signals

Nested message and enum types in FraudSignals.

ip_override_data

Nested message and enum types in IpOverrideData.

key

Nested message and enum types in Key.

recaptcha_enterprise_service_client

Generated client implementations.

risk_analysis

Nested message and enum types in RiskAnalysis.

sms_toll_fraud_verdict

Nested message and enum types in SmsTollFraudVerdict.

testing_options

Nested message and enum types in TestingOptions.

token_properties

Nested message and enum types in TokenProperties.

transaction_data

Nested message and enum types in TransactionData.

transaction_event

Nested message and enum types in TransactionEvent.

user_id

Nested message and enum types in UserId.

waf_settings

Nested message and enum types in WafSettings.

web_key_settings

Nested message and enum types in WebKeySettings.

Structs

AccountDefenderAssessment

Account defender risk assessment.

AccountVerificationInfo

Information about account verification, used for identity verification.

AddIpOverrideRequest

The AddIpOverride request message.

AddIpOverrideResponse

Response for AddIpOverride.

AndroidKeySettings

Settings specific to keys that can be used by Android apps.

AnnotateAssessmentRequest

The request message to annotate an Assessment.

AnnotateAssessmentResponse

Empty response for AnnotateAssessment.

AppleDeveloperId

Contains fields that are required to perform Apple-specific integrity checks.

Assessment

A reCAPTCHA Enterprise assessment resource.

AssessmentEnvironment

The environment creating the assessment. This describes your environment (the system invoking CreateAssessment), NOT the environment of your user.

ChallengeMetrics

Metrics related to challenges.

CreateAssessmentRequest

The create assessment request message.

CreateFirewallPolicyRequest

The create firewall policy request message.

CreateKeyRequest

The create key request message.

DeleteFirewallPolicyRequest

The delete firewall policy request message.

DeleteKeyRequest

The delete key request message.

EndpointVerificationInfo

Information about a verification endpoint that can be used for 2FA.

Event

The event being assessed.

ExpressKeySettings

Settings specific to keys that can be used for reCAPTCHA Express.

FirewallAction

An individual action. Each action represents what to do if a policy matches.

FirewallPolicy

A FirewallPolicy represents a single matching pattern and resulting actions to take.

FirewallPolicyAssessment

Policy config assessment.

FraudPreventionAssessment

Assessment for Fraud Prevention.

FraudSignals

Fraud signals describing users and cards involved in the transaction.

GetFirewallPolicyRequest

The get firewall policy request message.

GetKeyRequest

The get key request message.

GetMetricsRequest

The get metrics request message.

IosKeySettings

Settings specific to keys that can be used by iOS apps.

IpOverrideData

Information about the IP or IP range override.

Key

A key used to identify and configure applications (web and/or mobile) that use reCAPTCHA Enterprise.

ListFirewallPoliciesRequest

The list firewall policies request message.

ListFirewallPoliciesResponse

Response to request to list firewall policies belonging to a project.

ListKeysRequest

The list keys request message.

ListKeysResponse

Response to request to list keys in a project.

ListRelatedAccountGroupMembershipsRequest

The request message to list memberships in a related account group.

ListRelatedAccountGroupMembershipsResponse

The response to a ListRelatedAccountGroupMemberships call.

ListRelatedAccountGroupsRequest

The request message to list related account groups.

ListRelatedAccountGroupsResponse

The response to a ListRelatedAccountGroups call.

Metrics

Metrics for a single Key.

MigrateKeyRequest

The migrate key request message.

PhoneFraudAssessment

Assessment for Phone Fraud

PrivatePasswordLeakVerification

Private password leak verification info.

RelatedAccountGroup

A group of related accounts.

RelatedAccountGroupMembership

A membership in a group of related accounts.

ReorderFirewallPoliciesRequest

The reorder firewall policies request message.

ReorderFirewallPoliciesResponse

The reorder firewall policies response message.

RetrieveLegacySecretKeyRequest

The retrieve legacy secret key request message.

RetrieveLegacySecretKeyResponse

Secret key is used only in legacy reCAPTCHA. It must be used in a 3rd party integration with legacy reCAPTCHA.

RiskAnalysis

Risk analysis result for an event.

ScoreDistribution

Score distribution.

ScoreMetrics

Metrics related to scoring.

SearchRelatedAccountGroupMembershipsRequest

The request message to search related account group memberships.

SearchRelatedAccountGroupMembershipsResponse

The response to a SearchRelatedAccountGroupMemberships call.

SmsTollFraudVerdict

Information about SMS toll fraud.

TestingOptions

Options for user acceptance testing.

TokenProperties

Properties of the provided event token.

TransactionData

Transaction data associated with a payment protected by reCAPTCHA Enterprise.

TransactionEvent

Describes an event in the lifecycle of a payment transaction.

UpdateFirewallPolicyRequest

The update firewall policy request message.

UpdateKeyRequest

The update key request message.

UserId

An identifier associated with a user.

UserInfo

User information associated with a request protected by reCAPTCHA Enterprise.

WafSettings

Settings specific to keys that can be used for WAF (Web Application Firewall).

WebKeySettings

Settings specific to keys that can be used by websites.