Module google_api_proto::google::cloud::recaptchaenterprise::v1::firewall_action

Nested message and enum types in FirewallAction.

Structs

AllowAction

An allow action continues processing a request unimpeded.

BlockAction

A block action serves an HTTP error code a prevents the request from hitting the backend.

IncludeRecaptchaScriptAction

An include reCAPTCHA script action involves injecting reCAPTCHA JavaScript code into the HTML returned by the site backend. This reCAPTCHA script is tasked with collecting user signals on the requested web page, issuing tokens as a cookie within the site domain, and enabling their utilization in subsequent page requests.

RedirectAction

A redirect action returns a 307 (temporary redirect) response, pointing the user to a ReCaptcha interstitial page to attach a token.

SetHeaderAction

A set header action sets a header and forwards the request to the backend. This can be used to trigger custom protection implemented on the backend.

SubstituteAction

A substitute action transparently serves a different page than the one requested.

Enums

FirewallActionOneof