Struct google_api_proto::google::cloud::recaptchaenterprise::v1::recaptcha_enterprise_service_client::RecaptchaEnterpriseServiceClient

pub struct RecaptchaEnterpriseServiceClient<T> { /* private fields */ }

Service to determine the likelihood an event is legitimate.

Implementations

impl<T> RecaptchaEnterpriseServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> RecaptchaEnterpriseServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_assessment( &mut self, request: impl IntoRequest<CreateAssessmentRequest>, ) -> Result<Response<Assessment>, Status>

Creates an Assessment of the likelihood an event is legitimate.

pub async fn annotate_assessment( &mut self, request: impl IntoRequest<AnnotateAssessmentRequest>, ) -> Result<Response<AnnotateAssessmentResponse>, Status>

Annotates a previously created Assessment to provide additional information on whether the event turned out to be authentic or fraudulent.

pub async fn create_key( &mut self, request: impl IntoRequest<CreateKeyRequest>, ) -> Result<Response<Key>, Status>

Creates a new reCAPTCHA Enterprise key.

pub async fn list_keys( &mut self, request: impl IntoRequest<ListKeysRequest>, ) -> Result<Response<ListKeysResponse>, Status>

Returns the list of all keys that belong to a project.

pub async fn retrieve_legacy_secret_key( &mut self, request: impl IntoRequest<RetrieveLegacySecretKeyRequest>, ) -> Result<Response<RetrieveLegacySecretKeyResponse>, Status>

Returns the secret key related to the specified public key. You must use the legacy secret key only in a 3rd party integration with legacy reCAPTCHA.

pub async fn get_key( &mut self, request: impl IntoRequest<GetKeyRequest>, ) -> Result<Response<Key>, Status>

Returns the specified key.

pub async fn update_key( &mut self, request: impl IntoRequest<UpdateKeyRequest>, ) -> Result<Response<Key>, Status>

Updates the specified key.

pub async fn delete_key( &mut self, request: impl IntoRequest<DeleteKeyRequest>, ) -> Result<Response<()>, Status>

Deletes the specified key.

pub async fn migrate_key( &mut self, request: impl IntoRequest<MigrateKeyRequest>, ) -> Result<Response<Key>, Status>

Migrates an existing key from reCAPTCHA to reCAPTCHA Enterprise. Once a key is migrated, it can be used from either product. SiteVerify requests are billed as CreateAssessment calls. You must be authenticated as one of the current owners of the reCAPTCHA Key, and your user must have the reCAPTCHA Enterprise Admin IAM role in the destination project.

pub async fn add_ip_override( &mut self, request: impl IntoRequest<AddIpOverrideRequest>, ) -> Result<Response<AddIpOverrideResponse>, Status>

Adds an IP override to a key. The following restrictions hold:

The maximum number of IP overrides per key is 100.

For any conflict (such as IP already exists or IP part of an existing IP range), an error is returned.

pub async fn get_metrics( &mut self, request: impl IntoRequest<GetMetricsRequest>, ) -> Result<Response<Metrics>, Status>

Get some aggregated metrics for a Key. This data can be used to build dashboards.

pub async fn create_firewall_policy( &mut self, request: impl IntoRequest<CreateFirewallPolicyRequest>, ) -> Result<Response<FirewallPolicy>, Status>

Creates a new FirewallPolicy, specifying conditions at which reCAPTCHA Enterprise actions can be executed. A project may have a maximum of 1000 policies.

pub async fn list_firewall_policies( &mut self, request: impl IntoRequest<ListFirewallPoliciesRequest>, ) -> Result<Response<ListFirewallPoliciesResponse>, Status>

Returns the list of all firewall policies that belong to a project.

pub async fn get_firewall_policy( &mut self, request: impl IntoRequest<GetFirewallPolicyRequest>, ) -> Result<Response<FirewallPolicy>, Status>

Returns the specified firewall policy.

pub async fn update_firewall_policy( &mut self, request: impl IntoRequest<UpdateFirewallPolicyRequest>, ) -> Result<Response<FirewallPolicy>, Status>

Updates the specified firewall policy.

pub async fn delete_firewall_policy( &mut self, request: impl IntoRequest<DeleteFirewallPolicyRequest>, ) -> Result<Response<()>, Status>

Deletes the specified firewall policy.

pub async fn reorder_firewall_policies( &mut self, request: impl IntoRequest<ReorderFirewallPoliciesRequest>, ) -> Result<Response<ReorderFirewallPoliciesResponse>, Status>

Reorders all firewall policies.

pub async fn list_related_account_groups( &mut self, request: impl IntoRequest<ListRelatedAccountGroupsRequest>, ) -> Result<Response<ListRelatedAccountGroupsResponse>, Status>

List groups of related accounts.

pub async fn list_related_account_group_memberships( &mut self, request: impl IntoRequest<ListRelatedAccountGroupMembershipsRequest>, ) -> Result<Response<ListRelatedAccountGroupMembershipsResponse>, Status>

Get memberships in a group of related accounts.

pub async fn search_related_account_group_memberships( &mut self, request: impl IntoRequest<SearchRelatedAccountGroupMembershipsRequest>, ) -> Result<Response<SearchRelatedAccountGroupMembershipsResponse>, Status>

Search group memberships related to a given account.