Module google_api_proto::google::cloud::recommender::v1beta1

Modules

impact

Nested message and enum types in Impact.

insight

Nested message and enum types in Insight.

insight_state_info

Nested message and enum types in InsightStateInfo.

operation

Nested message and enum types in Operation.

recommendation

Nested message and enum types in Recommendation.

recommendation_state_info

Nested message and enum types in RecommendationStateInfo.

recommender_client

Generated client implementations.

value_matcher

Nested message and enum types in ValueMatcher.

Structs

CostProjection

Contains metadata about how much money a recommendation can save or incur.

GetInsightRequest

Request to the GetInsight method.

GetInsightTypeConfigRequest

Request for the GetInsightTypeConfig` method.

GetRecommendationRequest

Request to the GetRecommendation method.

GetRecommenderConfigRequest

Request for the GetRecommenderConfig` method.

Impact

Contains the impact a recommendation can have for a given category.

Insight

An insight along with the information used to derive the insight. The insight may have associated recommendations as well.

InsightStateInfo

Information related to insight state.

InsightType

The type of insight.

InsightTypeConfig

Configuration for an InsightType.

InsightTypeGenerationConfig

A configuration to customize the generation of insights. Eg, customizing the lookback period considered when generating a insight.

ListInsightTypesRequest

Request for the ListInsightTypes method.

ListInsightTypesResponse

Response for the ListInsightTypes method.

ListInsightsRequest

Request for the ListInsights method.

ListInsightsResponse

Response to the ListInsights method.

ListRecommendationsRequest

Request for the ListRecommendations method.

ListRecommendationsResponse

Response to the ListRecommendations method.

ListRecommendersRequest

Request for the ListRecommender method.

ListRecommendersResponse

Response for the ListRecommender method.

MarkInsightAcceptedRequest

Request for the MarkInsightAccepted method.

MarkRecommendationClaimedRequest

Request for the MarkRecommendationClaimed Method.

MarkRecommendationFailedRequest

Request for the MarkRecommendationFailed Method.

MarkRecommendationSucceededRequest

Request for the MarkRecommendationSucceeded Method.

Operation

Contains an operation for a resource loosely based on the JSON-PATCH format with support for:

OperationGroup

Group of operations that need to be performed atomically.

Recommendation

A recommendation along with a suggested action. E.g., a rightsizing recommendation for an underutilized VM, IAM role recommendations, etc

RecommendationContent

Contains what resources are changing and how they are changing.

RecommendationStateInfo

Information for state. Contains state and metadata.

RecommenderConfig

Configuration for a Recommender.

RecommenderGenerationConfig

A Configuration to customize the generation of recommendations. Eg, customizing the lookback period considered when generating a recommendation.

RecommenderType

The type of a recommender.

SecurityProjection

Contains various ways of describing the impact on Security.

SustainabilityProjection

Contains metadata about how much sustainability a recommendation can save or incur.

UpdateInsightTypeConfigRequest

Request for the UpdateInsightTypeConfig method.

UpdateRecommenderConfigRequest

Request for the UpdateRecommenderConfig method.

ValueMatcher

Contains various matching options for values for a GCP resource field.