Struct google_api_proto::google::cloud::recommender::v1beta1::recommender_client::RecommenderClient

pub struct RecommenderClient<T> { /* private fields */ }

Provides insights and recommendations for cloud customers for various categories like performance optimization, cost savings, reliability, feature discovery, etc. Insights and recommendations are generated automatically based on analysis of user resources, configuration and monitoring metrics.

Implementations

impl<T> RecommenderClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> RecommenderClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_insights( &mut self, request: impl IntoRequest<ListInsightsRequest>, ) -> Result<Response<ListInsightsResponse>, Status>

Lists insights for the specified Cloud Resource. Requires the recommender.*.list IAM permission for the specified insight type.

pub async fn get_insight( &mut self, request: impl IntoRequest<GetInsightRequest>, ) -> Result<Response<Insight>, Status>

Gets the requested insight. Requires the recommender.*.get IAM permission for the specified insight type.

pub async fn mark_insight_accepted( &mut self, request: impl IntoRequest<MarkInsightAcceptedRequest>, ) -> Result<Response<Insight>, Status>

Marks the Insight State as Accepted. Users can use this method to indicate to the Recommender API that they have applied some action based on the insight. This stops the insight content from being updated.

MarkInsightAccepted can be applied to insights in ACTIVE state. Requires the recommender.*.update IAM permission for the specified insight.

pub async fn list_recommendations( &mut self, request: impl IntoRequest<ListRecommendationsRequest>, ) -> Result<Response<ListRecommendationsResponse>, Status>

Lists recommendations for the specified Cloud Resource. Requires the recommender.*.list IAM permission for the specified recommender.

pub async fn get_recommendation( &mut self, request: impl IntoRequest<GetRecommendationRequest>, ) -> Result<Response<Recommendation>, Status>

Gets the requested recommendation. Requires the recommender.*.get IAM permission for the specified recommender.

pub async fn mark_recommendation_claimed( &mut self, request: impl IntoRequest<MarkRecommendationClaimedRequest>, ) -> Result<Response<Recommendation>, Status>

Marks the Recommendation State as Claimed. Users can use this method to indicate to the Recommender API that they are starting to apply the recommendation themselves. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state.

MarkRecommendationClaimed can be applied to recommendations in CLAIMED or ACTIVE state.

Requires the recommender.*.update IAM permission for the specified recommender.

pub async fn mark_recommendation_succeeded( &mut self, request: impl IntoRequest<MarkRecommendationSucceededRequest>, ) -> Result<Response<Recommendation>, Status>

Marks the Recommendation State as Succeeded. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation was successful. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state.

MarkRecommendationSucceeded can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state.

Requires the recommender.*.update IAM permission for the specified recommender.

pub async fn mark_recommendation_failed( &mut self, request: impl IntoRequest<MarkRecommendationFailedRequest>, ) -> Result<Response<Recommendation>, Status>

Marks the Recommendation State as Failed. Users can use this method to indicate to the Recommender API that they have applied the recommendation themselves, and the operation failed. This stops the recommendation content from being updated. Associated insights are frozen and placed in the ACCEPTED state.

MarkRecommendationFailed can be applied to recommendations in ACTIVE, CLAIMED, SUCCEEDED, or FAILED state.

Requires the recommender.*.update IAM permission for the specified recommender.

pub async fn get_recommender_config( &mut self, request: impl IntoRequest<GetRecommenderConfigRequest>, ) -> Result<Response<RecommenderConfig>, Status>

Gets the requested Recommender Config. There is only one instance of the config for each Recommender.

pub async fn update_recommender_config( &mut self, request: impl IntoRequest<UpdateRecommenderConfigRequest>, ) -> Result<Response<RecommenderConfig>, Status>

Updates a Recommender Config. This will create a new revision of the config.

pub async fn get_insight_type_config( &mut self, request: impl IntoRequest<GetInsightTypeConfigRequest>, ) -> Result<Response<InsightTypeConfig>, Status>

Gets the requested InsightTypeConfig. There is only one instance of the config for each InsightType.

pub async fn update_insight_type_config( &mut self, request: impl IntoRequest<UpdateInsightTypeConfigRequest>, ) -> Result<Response<InsightTypeConfig>, Status>

Updates an InsightTypeConfig change. This will create a new revision of the config.

pub async fn list_recommenders( &mut self, request: impl IntoRequest<ListRecommendersRequest>, ) -> Result<Response<ListRecommendersResponse>, Status>

Lists all available Recommenders. No IAM permissions are required.

pub async fn list_insight_types( &mut self, request: impl IntoRequest<ListInsightTypesRequest>, ) -> Result<Response<ListInsightTypesResponse>, Status>

Lists available InsightTypes. No IAM permissions are required.