Struct google_api_proto::google::cloud::redis::v1beta1::cloud_redis_client::CloudRedisClient

pub struct CloudRedisClient<T> { /* private fields */ }

Configures and manages Cloud Memorystore for Redis instances

Google Cloud Memorystore for Redis v1beta1

The redis.googleapis.com service implements the Google Cloud Memorystore for Redis API and defines the following resource model for managing Redis instances:

The service works with a collection of cloud projects, named: /projects/*

Each project has a collection of available locations, named: /locations/*

Each location has a collection of Redis instances, named: /instances/*

As such, Redis instances are resources of the form: /projects/{project_id}/locations/{location_id}/instances/{instance_id}

Note that location_id must be referring to a GCP region; for example:

projects/redpepper-1290/locations/us-central1/instances/my-redis

Implementations

impl<T> CloudRedisClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> CloudRedisClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_instances( &mut self, request: impl IntoRequest<ListInstancesRequest> ) -> Result<Response<ListInstancesResponse>, Status>

Lists all Redis instances owned by a project in either the specified location (region) or all locations.

The location should have the following format:

projects/{project_id}/locations/{location_id}

If location_id is specified as - (wildcard), then all regions available to the project are queried, and the results are aggregated.

pub async fn get_instance( &mut self, request: impl IntoRequest<GetInstanceRequest> ) -> Result<Response<Instance>, Status>

Gets the details of a specific Redis instance.

pub async fn get_instance_auth_string( &mut self, request: impl IntoRequest<GetInstanceAuthStringRequest> ) -> Result<Response<InstanceAuthString>, Status>

Gets the AUTH string for a Redis instance. If AUTH is not enabled for the instance the response will be empty. This information is not included in the details returned to GetInstance.

pub async fn create_instance( &mut self, request: impl IntoRequest<CreateInstanceRequest> ) -> Result<Response<Operation>, Status>

Creates a Redis instance based on the specified tier and memory size.

By default, the instance is accessible from the project’s default network.

The creation is executed asynchronously and callers may check the returned operation to track its progress. Once the operation is completed the Redis instance will be fully functional. The completed longrunning.Operation will contain the new instance object in the response field.

The returned operation is automatically deleted after a few hours, so there is no need to call DeleteOperation.

pub async fn update_instance( &mut self, request: impl IntoRequest<UpdateInstanceRequest> ) -> Result<Response<Operation>, Status>

Updates the metadata and configuration of a specific Redis instance.

Completed longrunning.Operation will contain the new instance object in the response field. The returned operation is automatically deleted after a few hours, so there is no need to call DeleteOperation.

pub async fn upgrade_instance( &mut self, request: impl IntoRequest<UpgradeInstanceRequest> ) -> Result<Response<Operation>, Status>

Upgrades Redis instance to the newer Redis version specified in the request.

pub async fn import_instance( &mut self, request: impl IntoRequest<ImportInstanceRequest> ) -> Result<Response<Operation>, Status>

Import a Redis RDB snapshot file from Cloud Storage into a Redis instance.

Redis may stop serving during this operation. Instance state will be IMPORTING for entire operation. When complete, the instance will contain only data from the imported file.

The returned operation is automatically deleted after a few hours, so there is no need to call DeleteOperation.

pub async fn export_instance( &mut self, request: impl IntoRequest<ExportInstanceRequest> ) -> Result<Response<Operation>, Status>

Export Redis instance data into a Redis RDB format file in Cloud Storage.

Redis will continue serving during this operation.

The returned operation is automatically deleted after a few hours, so there is no need to call DeleteOperation.

pub async fn failover_instance( &mut self, request: impl IntoRequest<FailoverInstanceRequest> ) -> Result<Response<Operation>, Status>

Initiates a failover of the primary node to current replica node for a specific STANDARD tier Cloud Memorystore for Redis instance.

pub async fn delete_instance( &mut self, request: impl IntoRequest<DeleteInstanceRequest> ) -> Result<Response<Operation>, Status>

Deletes a specific Redis instance. Instance stops serving and data is deleted.

pub async fn reschedule_maintenance( &mut self, request: impl IntoRequest<RescheduleMaintenanceRequest> ) -> Result<Response<Operation>, Status>

Reschedule maintenance for a given instance in a given project and location.