Struct google_api_proto::google::cloud::redis::v1beta1::Instance

pub struct Instance {

    pub name: String,
    pub display_name: String,
    pub labels: BTreeMap<String, String>,
    pub location_id: String,
    pub alternative_location_id: String,
    pub redis_version: String,
    pub reserved_ip_range: String,
    pub secondary_ip_range: String,
    pub host: String,
    pub port: i32,
    pub current_location_id: String,
    pub create_time: Option<Timestamp>,
    pub state: i32,
    pub status_message: String,
    pub redis_configs: BTreeMap<String, String>,
    pub tier: i32,
    pub memory_size_gb: i32,
    pub authorized_network: String,
    pub persistence_iam_identity: String,
    pub connect_mode: i32,
    pub auth_enabled: bool,
    pub server_ca_certs: Vec<TlsCertificate>,
    pub transit_encryption_mode: i32,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub maintenance_schedule: Option<MaintenanceSchedule>,
    pub replica_count: i32,
    pub nodes: Vec<NodeInfo>,
    pub read_endpoint: String,
    pub read_endpoint_port: i32,
    pub read_replicas_mode: i32,
    pub persistence_config: Option<PersistenceConfig>,

}

A Memorystore for Redis instance.

Fields
name: String

Required. Unique name of the resource in this scope including project and location using the form: projects/{project_id}/locations/{location_id}/instances/{instance_id}

Note: Redis instances are managed and addressed at regional level so location_id here refers to a GCP region; however, users may choose which specific zone (or collection of zones for cross-zone instances) an instance should be provisioned in. Refer to [location_id][google.cloud.redis.v1beta1.Instance.location_id] and [alternative_location_id][google.cloud.redis.v1beta1.Instance.alternative_location_id] fields for more details.

display_name: String

An arbitrary and optional user-provided name for the instance.

labels: BTreeMap<String, String>

Resource labels to represent user provided metadata

location_id: String

Optional. The zone where the instance will be provisioned. If not provided, the service will choose a zone from the specified region for the instance. For standard tier, additional nodes will be added across multiple zones for protection against zonal failures. If specified, at least one node will be provisioned in this zone.

alternative_location_id: String

Optional. If specified, at least one node will be provisioned in this zone in addition to the zone specified in location_id. Only applicable to standard tier. If provided, it must be a different zone from the one provided in [location_id]. Additional nodes beyond the first 2 will be placed in zones selected by the service.

redis_version: String

Optional. The version of Redis software. If not provided, latest supported version will be used. Currently, the supported values are:

REDIS_3_2 for Redis 3.2 compatibility

REDIS_4_0 for Redis 4.0 compatibility (default)

REDIS_5_0 for Redis 5.0 compatibility

REDIS_6_X for Redis 6.x compatibility

reserved_ip_range: String

Optional. For DIRECT_PEERING mode, the CIDR range of internal addresses that are reserved for this instance. Range must be unique and non-overlapping with existing subnets in an authorized network. For PRIVATE_SERVICE_ACCESS mode, the name of one allocated IP address ranges associated with this private service access connection. If not provided, the service will choose an unused /29 block, for example, 10.0.0.0/29 or 192.168.0.0/29. For READ_REPLICAS_ENABLED the default block size is /28.

secondary_ip_range: String

Optional. Additional IP range for node placement. Required when enabling read replicas on an existing instance. For DIRECT_PEERING mode value must be a CIDR range of size /28, or “auto”. For PRIVATE_SERVICE_ACCESS mode value must be the name of an allocated address range associated with the private service access connection, or “auto”.

host: String

Output only. Hostname or IP address of the exposed Redis endpoint used by clients to connect to the service.

port: i32

Output only. The port number of the exposed Redis endpoint.

current_location_id: String

Output only. The current zone where the Redis primary node is located. In basic tier, this will always be the same as [location_id]. In standard tier, this can be the zone of any node in the instance.

create_time: Option<Timestamp>

Output only. The time the instance was created.

state: i32

Output only. The current state of this instance.

status_message: String

Output only. Additional information about the current status of this instance, if available.

redis_configs: BTreeMap<String, String>

Optional. Redis configuration parameters, according to http://redis.io/topics/config. Currently, the only supported parameters are:

Redis version 3.2 and newer:

maxmemory-policy

notify-keyspace-events

Redis version 4.0 and newer:

activedefrag

lfu-decay-time

lfu-log-factor

maxmemory-gb

Redis version 5.0 and newer:

stream-node-max-bytes

stream-node-max-entries

tier: i32

Required. The service tier of the instance.

memory_size_gb: i32

Required. Redis memory size in GiB.

authorized_network: String

Optional. The full name of the Google Compute Engine network to which the instance is connected. If left unspecified, the default network will be used.

persistence_iam_identity: String

Output only. Cloud IAM identity used by import / export operations to transfer data to/from Cloud Storage. Format is “serviceAccount:<service_account_email>”. The value may change over time for a given instance so should be checked before each import/export operation.

connect_mode: i32

Optional. The network connect mode of the Redis instance. If not provided, the connect mode defaults to DIRECT_PEERING.

auth_enabled: bool

Optional. Indicates whether OSS Redis AUTH is enabled for the instance. If set to “true” AUTH is enabled on the instance. Default value is “false” meaning AUTH is disabled.

server_ca_certs: Vec<TlsCertificate>

Output only. List of server CA certificates for the instance.

transit_encryption_mode: i32

Optional. The TLS mode of the Redis instance. If not provided, TLS is disabled for the instance.

maintenance_policy: Option<MaintenancePolicy>

Optional. The maintenance policy for the instance. If not provided, maintenance events can be performed at any time.

maintenance_schedule: Option<MaintenanceSchedule>

Output only. Date and time of upcoming maintenance events which have been scheduled.

replica_count: i32

Optional. The number of replica nodes. The valid range for the Standard Tier with read replicas enabled is [1-5] and defaults to 2. If read replicas are not enabled for a Standard Tier instance, the only valid value is 1 and the default is 1. The valid value for basic tier is 0 and the default is also 0.

nodes: Vec<NodeInfo>

Output only. Info per node.

read_endpoint: String

Output only. Hostname or IP address of the exposed readonly Redis endpoint. Standard tier only. Targets all healthy replica nodes in instance. Replication is asynchronous and replica nodes will exhibit some lag behind the primary. Write requests must target ‘host’.

read_endpoint_port: i32

Output only. The port number of the exposed readonly redis endpoint. Standard tier only. Write requests should target ‘port’.

read_replicas_mode: i32

Optional. Read replicas mode for the instance. Defaults to READ_REPLICAS_DISABLED.

persistence_config: Option<PersistenceConfig>

Optional. Persistence configuration parameters

Implementations

impl Instance

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.

pub fn tier(&self) -> Tier

Returns the enum value of tier, or the default if the field is set to an invalid enum value.

pub fn set_tier(&mut self, value: Tier)

Sets tier to the provided enum value.

pub fn connect_mode(&self) -> ConnectMode

Returns the enum value of connect_mode, or the default if the field is set to an invalid enum value.

pub fn set_connect_mode(&mut self, value: ConnectMode)

Sets connect_mode to the provided enum value.

pub fn transit_encryption_mode(&self) -> TransitEncryptionMode

Returns the enum value of transit_encryption_mode, or the default if the field is set to an invalid enum value.

pub fn set_transit_encryption_mode(&mut self, value: TransitEncryptionMode)

Sets transit_encryption_mode to the provided enum value.

pub fn read_replicas_mode(&self) -> ReadReplicasMode

Returns the enum value of read_replicas_mode, or the default if the field is set to an invalid enum value.

pub fn set_read_replicas_mode(&mut self, value: ReadReplicasMode)

Sets read_replicas_mode to the provided enum value.