Module google_api_proto::google::cloud::resourcemanager::v3

Modules

folder

Nested message and enum types in Folder.

folders_client

Generated client implementations.

organization

Nested message and enum types in Organization.

organizations_client

Generated client implementations.

project

Nested message and enum types in Project.

projects_client

Generated client implementations.

tag_bindings_client

Generated client implementations.

tag_holds_client

Generated client implementations.

tag_keys_client

Generated client implementations.

tag_values_client

Generated client implementations.

Structs

CreateFolderMetadata

Metadata pertaining to the Folder creation process.

CreateFolderRequest

The CreateFolder request message.

CreateProjectMetadata

A status object which is used as the metadata field for the Operation returned by CreateProject. It provides insight for when significant phases of Project creation have completed.

CreateProjectRequest

The request sent to the [CreateProject][google.cloud.resourcemanager.v3.Projects.CreateProject] method.

CreateTagBindingMetadata

Runtime operation information for creating a TagValue.

CreateTagBindingRequest

The request message to create a TagBinding.

CreateTagHoldMetadata

Runtime operation information for creating a TagHold. (– The metadata is currently empty, but may include information in the future. –)

CreateTagHoldRequest

The request message to create a TagHold.

CreateTagKeyMetadata

Runtime operation information for creating a TagKey.

CreateTagKeyRequest

The request message for creating a TagKey.

CreateTagValueMetadata

Runtime operation information for creating a TagValue.

CreateTagValueRequest

The request message for creating a TagValue.

DeleteFolderMetadata

A status object which is used as the metadata field for the Operation returned by DeleteFolder.

DeleteFolderRequest

The DeleteFolder request message.

DeleteOrganizationMetadata

A status object which is used as the metadata field for the operation returned by DeleteOrganization.

DeleteProjectMetadata

A status object which is used as the metadata field for the Operation returned by DeleteProject.

DeleteProjectRequest

[DeleteProject][google.cloud.resourcemanager.v3.Projects.DeleteProject] method.

DeleteTagBindingMetadata

Runtime operation information for deleting a TagBinding.

DeleteTagBindingRequest

The request message to delete a TagBinding.

DeleteTagHoldMetadata

Runtime operation information for deleting a TagHold. (– The metadata is currently empty, but may include information in the future. –)

DeleteTagHoldRequest

The request message to delete a TagHold.

DeleteTagKeyMetadata

Runtime operation information for deleting a TagKey.

DeleteTagKeyRequest

The request message for deleting a TagKey.

DeleteTagValueMetadata

Runtime operation information for deleting a TagValue.

DeleteTagValueRequest

The request message for deleting a TagValue.

EffectiveTag

An EffectiveTag represents a tag that applies to a resource during policy evaluation. Tags can be either directly bound to a resource or inherited from its ancestor. EffectiveTag contains the name and namespaced_name of the tag value and tag key, with additional fields of inherited to indicate the inheritance status of the effective tag.

Folder

A folder in an organization’s resource hierarchy, used to organize that organization’s resources.

GetFolderRequest

The GetFolder request message.

GetNamespacedTagKeyRequest

The request message for getting a TagKey by its namespaced name.

GetNamespacedTagValueRequest

The request message for getting a TagValue by its namespaced name.

GetOrganizationRequest

The request sent to the GetOrganization method. The name field is required. organization_id is no longer accepted.

GetProjectRequest

The request sent to the [GetProject][google.cloud.resourcemanager.v3.Projects.GetProject] method.

GetTagKeyRequest

The request message for getting a TagKey.

GetTagValueRequest

The request message for getting a TagValue.

ListEffectiveTagsRequest

The request message to ListEffectiveTags

ListEffectiveTagsResponse

The response of ListEffectiveTags.

ListFoldersRequest

The ListFolders request message.

ListFoldersResponse

The ListFolders response message.

ListProjectsRequest

The request sent to the [ListProjects][google.cloud.resourcemanager.v3.Projects.ListProjects] method.

ListProjectsResponse

A page of the response received from the [ListProjects][google.cloud.resourcemanager.v3.Projects.ListProjects] method.

ListTagBindingsRequest

The request message to list all TagBindings for a parent.

ListTagBindingsResponse

The ListTagBindings response.

ListTagHoldsRequest

The request message for listing the TagHolds under a TagValue.

ListTagHoldsResponse

The ListTagHolds response.

ListTagKeysRequest

The request message for listing all TagKeys under a parent resource.

ListTagKeysResponse

The ListTagKeys response message.

ListTagValuesRequest

The request message for listing TagValues for the specified TagKey. Resource name for TagKey, parent of the TagValues to be listed, in the format tagKeys/123.

ListTagValuesResponse

The ListTagValues response.

MoveFolderMetadata

Metadata pertaining to the folder move process.

MoveFolderRequest

The MoveFolder request message.

MoveProjectMetadata

A status object which is used as the metadata field for the Operation returned by MoveProject.

MoveProjectRequest

The request sent to [MoveProject][google.cloud.resourcemanager.v3.Projects.MoveProject] method.

Organization

The root node in the resource hierarchy to which a particular entity’s (a company, for example) resources belong.

Project

A project is a high-level Google Cloud entity. It is a container for ACLs, APIs, App Engine Apps, VMs, and other Google Cloud Platform resources.

SearchFoldersRequest

The request message for searching folders.

SearchFoldersResponse

The response message for searching folders.

SearchOrganizationsRequest

The request sent to the SearchOrganizations method.

SearchOrganizationsResponse

The response returned from the SearchOrganizations method.

SearchProjectsRequest

The request sent to the [SearchProjects][google.cloud.resourcemanager.v3.Projects.SearchProjects] method.

SearchProjectsResponse

A page of the response received from the [SearchProjects][google.cloud.resourcemanager.v3.Projects.SearchProjects] method.

TagBinding

A TagBinding represents a connection between a TagValue and a cloud resource Once a TagBinding is created, the TagValue is applied to all the descendants of the Google Cloud resource.

TagHold

A TagHold represents the use of a TagValue that is not captured by TagBindings. If a TagValue has any TagHolds, deletion will be blocked. This resource is intended to be created in the same cloud location as the holder.

TagKey

A TagKey, used to group a set of TagValues.

TagValue

A TagValue is a child of a particular TagKey. This is used to group cloud resources for the purpose of controlling them using policies.

UndeleteFolderMetadata

A status object which is used as the metadata field for the Operation returned by UndeleteFolder.

UndeleteFolderRequest

The UndeleteFolder request message.

UndeleteOrganizationMetadata

A status object which is used as the metadata field for the Operation returned by UndeleteOrganization.

UndeleteProjectMetadata

A status object which is used as the metadata field for the Operation returned by UndeleteProject.

UndeleteProjectRequest

The request sent to the [UndeleteProject] [google.cloud.resourcemanager.v3.Projects.UndeleteProject] method.

UpdateFolderMetadata

A status object which is used as the metadata field for the Operation returned by UpdateFolder.

UpdateFolderRequest

The request sent to the [UpdateFolder][google.cloud.resourcemanager.v3.Folder.UpdateFolder] method.

UpdateProjectMetadata

A status object which is used as the metadata field for the Operation returned by UpdateProject.

UpdateProjectRequest

The request sent to the [UpdateProject][google.cloud.resourcemanager.v3.Projects.UpdateProject] method.

UpdateTagKeyMetadata

Runtime operation information for updating a TagKey.

UpdateTagKeyRequest

The request message for updating a TagKey.

UpdateTagValueMetadata

Runtime operation information for updating a TagValue.

UpdateTagValueRequest

The request message for updating a TagValue.

Enums

Purpose

A purpose for each policy engine requiring such an integration. A single policy engine may have multiple purposes defined, however a TagKey may only specify a single purpose.