Module google_api_proto::google::cloud::retail::v2::rule

Nested message and enum types in Rule.

Modules

force_return_facet_action

Nested message and enum types in ForceReturnFacetAction.

Structs

BoostAction

A boost action to apply to results matching condition specified above.

DoNotAssociateAction

Prevents query_term from being associated with specified terms during search. Example: Don’t associate “gShoe” and “cheap”.

FilterAction

Rule Condition:No [Condition.query_terms][google.cloud.retail.v2.Condition.query_terms] provided is a global match.1 or more [Condition.query_terms][google.cloud.retail.v2.Condition.query_terms] provided are combined with OR operator.Action Input: The request query and filter that are applied to the retrieved products, in addition to any filters already provided with the SearchRequest. The AND operator is used to combine the query’s existing filters with the filter rule(s). NOTE: May result in 0 results when filters conflict.Action Result: Filters the returned objects to be ONLY those that passed the filter.

ForceReturnFacetAction

Force returns an attribute/facet in the request around a certain position or above.

IgnoreAction

Prevents a term in the query from being used in search. Example: Don’t search for “shoddy”.

OnewaySynonymsAction

Maps a set of terms to a set of synonyms. Set of synonyms will be treated as synonyms of each query term only. query_terms will not be treated as synonyms of each other. Example: “sneakers” will use a synonym of “shoes”. “shoes” will not use a synonym of “sneakers”.

RedirectAction

Redirects a shopper to a specific page.

RemoveFacetAction

Removes an attribute/facet in the request if is present.

ReplacementAction

Replaces a term in the query. Multiple replacement candidates can be specified. All query_terms will be replaced with the replacement term. Example: Replace “gShoe” with “google shoe”.

TwowaySynonymsAction

Creates a set of terms that will be treated as synonyms of each other. Example: synonyms of “sneakers” and “shoes”:

Enums

Action

An action must be provided.