Struct google_api_proto::google::cloud::retail::v2::Control

pub struct Control {
    pub name: String,
    pub display_name: String,
    pub associated_serving_config_ids: Vec<String>,
    pub solution_types: Vec<i32>,
    pub search_solution_use_case: Vec<i32>,
    pub control: Option<Control>,
}

Configures dynamic metadata that can be linked to a [ServingConfig][google.cloud.retail.v2.ServingConfig] and affect search or recommendation results at serving time.

Fields
name: String

Immutable. Fully qualified name projects/*/locations/global/catalogs/*/controls/*

display_name: String

Required. The human readable control display name. Used in Retail UI.

This field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is thrown.

associated_serving_config_ids: Vec<String>

Output only. List of [serving config][google.cloud.retail.v2.ServingConfig] ids that are associated with this control in the same [Catalog][google.cloud.retail.v2.Catalog].

Note the association is managed via the [ServingConfig][google.cloud.retail.v2.ServingConfig], this is an output only denormalized view.

solution_types: Vec<i32>

Required. Immutable. The solution types that the control is used for. Currently we support setting only one type of solution at creation time.

Only SOLUTION_TYPE_SEARCH value is supported at the moment. If no solution type is provided at creation time, will default to [SOLUTION_TYPE_SEARCH][google.cloud.retail.v2.SolutionType.SOLUTION_TYPE_SEARCH].

search_solution_use_case: Vec<i32>

Specifies the use case for the control. Affects what condition fields can be set. Only settable by search controls. Will default to [SEARCH_SOLUTION_USE_CASE_SEARCH][google.cloud.retail.v2.SearchSolutionUseCase.SEARCH_SOLUTION_USE_CASE_SEARCH] if not specified. Currently only allow one search_solution_use_case per control.

control: Option<Control>

The behavior/type of the control

A behavior/type must be specified on creation. Type cannot be changed once specified (e.g. A Rule control will always be a Rule control.). An INVALID_ARGUMENT will be returned if either condition is violated.

Implementations

impl Control

pub fn solution_types( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SolutionType>>

Returns an iterator which yields the valid enum values contained in solution_types.

pub fn push_solution_types(&mut self, value: SolutionType)

Appends the provided enum value to solution_types.

pub fn search_solution_use_case( &self, ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SearchSolutionUseCase>>

Returns an iterator which yields the valid enum values contained in search_solution_use_case.

pub fn push_search_solution_use_case(&mut self, value: SearchSolutionUseCase)

Appends the provided enum value to search_solution_use_case.