Struct google_api_proto::google::cloud::retail::v2::UserEvent

pub struct UserEvent {

    pub event_type: String,
    pub visitor_id: String,
    pub session_id: String,
    pub event_time: Option<Timestamp>,
    pub experiment_ids: Vec<String>,
    pub attribution_token: String,
    pub product_details: Vec<ProductDetail>,
    pub completion_detail: Option<CompletionDetail>,
    pub attributes: BTreeMap<String, CustomAttribute>,
    pub cart_id: String,
    pub purchase_transaction: Option<PurchaseTransaction>,
    pub search_query: String,
    pub filter: String,
    pub order_by: String,
    pub offset: i32,
    pub page_categories: Vec<String>,
    pub user_info: Option<UserInfo>,
    pub uri: String,
    pub referrer_uri: String,
    pub page_view_id: String,
    pub entity: String,

}

UserEvent captures all metadata information Retail API needs to know about how end users interact with customers’ website.

Fields
event_type: String

Required. User event type. Allowed values are:

add-to-cart: Products being added to cart.

remove-from-cart: Products being removed from cart.

category-page-view: Special pages such as sale or promotion pages viewed.

detail-page-view: Products detail page viewed.

home-page-view: Homepage viewed.

promotion-offered: Promotion is offered to a user.

promotion-not-offered: Promotion is not offered to a user.

purchase-complete: User finishing a purchase.

search: Product search.

shopping-cart-page-view: User viewing a shopping cart.

visitor_id: String

Required. A unique identifier for tracking visitors.

For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor log in/out of the website.

Don’t set the field to the same fixed ID for different users. This mixes the event history of those users together, which results in degraded model quality.

The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.

The field should not contain PII or user-data. We recommend to use Google Analytics Client ID for this field.

session_id: String

A unique identifier for tracking a visitor session with a length limit of 128 bytes. A session is an aggregation of an end user behavior in a time span.

A general guideline to populate the sesion_id:

If user has no activity for 30 min, a new session_id should be assigned.

The session_id should be unique across users, suggest use uuid or add visitor_id as prefix.

event_time: Option<Timestamp>

Only required for [UserEventService.ImportUserEvents][google.cloud.retail.v2.UserEventService.ImportUserEvents] method. Timestamp of when the user event happened.

experiment_ids: Vec<String>

A list of identifiers for the independent experiment groups this user event belongs to. This is used to distinguish between user events associated with different experiment setups (e.g. using Retail API, using different recommendation models).

attribution_token: String

Highly recommended for user events that are the result of [PredictionService.Predict][google.cloud.retail.v2.PredictionService.Predict]. This field enables accurate attribution of recommendation model performance.

The value must be a valid [PredictResponse.attribution_token][google.cloud.retail.v2.PredictResponse.attribution_token] for user events that are the result of [PredictionService.Predict][google.cloud.retail.v2.PredictionService.Predict]. The value must be a valid [SearchResponse.attribution_token][google.cloud.retail.v2.SearchResponse.attribution_token] for user events that are the result of [SearchService.Search][google.cloud.retail.v2.SearchService.Search].

This token enables us to accurately attribute page view or purchase back to the event and the particular predict response containing this clicked/purchased product. If user clicks on product K in the recommendation results, pass [PredictResponse.attribution_token][google.cloud.retail.v2.PredictResponse.attribution_token] as a URL parameter to product K’s page. When recording events on product K’s page, log the [PredictResponse.attribution_token][google.cloud.retail.v2.PredictResponse.attribution_token] to this field.

product_details: Vec<ProductDetail>

The main product details related to the event.

This field is optional except for the following event types:

add-to-cart

detail-page-view

purchase-complete

In a search event, this field represents the products returned to the end user on the current page (the end user may have not finished browsing the whole page yet). When a new page is returned to the end user, after pagination/filtering/ordering even for the same query, a new search event with different [product_details][google.cloud.retail.v2.UserEvent.product_details] is desired. The end user may have not finished browsing the whole page yet.

completion_detail: Option<CompletionDetail>

The main auto-completion details related to the event.

This field should be set for search event when autocomplete function is enabled and the user clicks a suggestion for search.

attributes: BTreeMap<String, CustomAttribute>

Extra user event features to include in the recommendation model.

If you provide custom attributes for ingested user events, also include them in the user events that you associate with prediction requests. Custom attribute formatting must be consistent between imported events and events provided with prediction requests. This lets the Retail API use those custom attributes when training models and serving predictions, which helps improve recommendation quality.

This field needs to pass all below criteria, otherwise an INVALID_ARGUMENT error is returned:

The key must be a UTF-8 encoded string with a length limit of 5,000 characters.

For text attributes, at most 400 values are allowed. Empty values are not allowed. Each value must be a UTF-8 encoded string with a length limit of 256 characters.

For number attributes, at most 400 values are allowed.

For product recommendations, an example of extra user information is traffic_channel, which is how a user arrives at the site. Users can arrive at the site by coming to the site directly, coming through Google search, or in other ways.

cart_id: String

The ID or name of the associated shopping cart. This ID is used to associate multiple items added or present in the cart before purchase.

This can only be set for add-to-cart, purchase-complete, or shopping-cart-page-view events.

purchase_transaction: Option<PurchaseTransaction>

A transaction represents the entire purchase transaction.

Required for purchase-complete events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

search_query: String

The user’s search query.

See [SearchRequest.query][google.cloud.retail.v2.SearchRequest.query] for definition.

The value must be a UTF-8 encoded string with a length limit of 5,000 characters. Otherwise, an INVALID_ARGUMENT error is returned.

At least one of [search_query][google.cloud.retail.v2.UserEvent.search_query] or [page_categories][google.cloud.retail.v2.UserEvent.page_categories] is required for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

filter: String

The filter syntax consists of an expression language for constructing a predicate from one or more fields of the products being filtered.

See [SearchRequest.filter][google.cloud.retail.v2.SearchRequest.filter] for definition and syntax.

The value must be a UTF-8 encoded string with a length limit of 1,000 characters. Otherwise, an INVALID_ARGUMENT error is returned.

order_by: String

The order in which products are returned.

See [SearchRequest.order_by][google.cloud.retail.v2.SearchRequest.order_by] for definition and syntax.

The value must be a UTF-8 encoded string with a length limit of 1,000 characters. Otherwise, an INVALID_ARGUMENT error is returned.

This can only be set for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

offset: i32

An integer that specifies the current offset for pagination (the 0-indexed starting location, amongst the products deemed by the API as relevant).

See [SearchRequest.offset][google.cloud.retail.v2.SearchRequest.offset] for definition.

If this field is negative, an INVALID_ARGUMENT is returned.

This can only be set for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

page_categories: Vec<String>

The categories associated with a category page.

To represent full path of category, use ‘>’ sign to separate different hierarchies. If ‘>’ is part of the category name, replace it with other character(s).

Category pages include special pages such as sales or promotions. For instance, a special sale page may have the category hierarchy: “pageCategories” : [“Sales > 2017 Black Friday Deals”].

Required for category-page-view events. At least one of [search_query][google.cloud.retail.v2.UserEvent.search_query] or [page_categories][google.cloud.retail.v2.UserEvent.page_categories] is required for search events. Other event types should not set this field. Otherwise, an INVALID_ARGUMENT error is returned.

user_info: Option<UserInfo>

User information.

uri: String

Complete URL (window.location.href) of the user’s current page.

When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically. Maximum length 5,000 characters.

referrer_uri: String

The referrer URL of the current page.

When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically.

page_view_id: String

A unique ID of a web page view.

This should be kept the same for all user events triggered from the same pageview. For example, an item detail page view could trigger multiple events as the user is browsing the page. The pageViewId property should be kept the same for all these events so that they can be grouped together properly.

When using the client side event reporting with JavaScript pixel and Google Tag Manager, this value is filled in automatically.

entity: String

The entity for customers that may run multiple different entities, domains, sites or regions, for example, Google US, Google Ads, Waymo, google.com, youtube.com, etc. We recommend that you set this field to get better per-entity search, completion, and prediction results.