Module google_api_proto::google::cloud::retail::v2alpha::catalog_attribute::facet_config

Nested message and enum types in FacetConfig.

Structs

IgnoredFacetValues

[Facet values][google.cloud.retail.v2alpha.SearchResponse.Facet.values] to ignore on [facets][google.cloud.retail.v2alpha.SearchResponse.Facet] during the specified time range for the given [SearchResponse.Facet.key][google.cloud.retail.v2alpha.SearchResponse.Facet.key] attribute.

MergedFacet

The current facet key (i.e. attribute config) maps into the [merged_facet_key][google.cloud.retail.v2alpha.CatalogAttribute.FacetConfig.MergedFacet.merged_facet_key]. A facet key can have at most one child. The current facet key and the merged facet key need both to be textual custom attributes or both numerical custom attributes (same type).

MergedFacetValue

Replaces a set of textual facet values by the same (possibly different) merged facet value. Each facet value should appear at most once as a value per [CatalogAttribute][google.cloud.retail.v2alpha.CatalogAttribute]. This feature is available only for textual custom attributes.

RerankConfig

Options to rerank based on facet values engaged by the user for the current key. That key needs to be a custom textual key and facetable. To use this control, you also need to pass all the facet keys engaged by the user in the request using the field [SearchRequest.FacetSpec]. In particular, if you don’t pass the facet keys engaged that you want to rerank on, this control won’t be effective. Moreover, to obtain better results, the facet values that you want to rerank on should be close to English (ideally made of words, underscores, and spaces).