Struct google_api_proto::google::cloud::retail::v2alpha::SearchRequest

pub struct SearchRequest {

    pub placement: String,
    pub branch: String,
    pub query: String,
    pub visitor_id: String,
    pub user_info: Option<UserInfo>,
    pub page_size: i32,
    pub page_token: String,
    pub offset: i32,
    pub filter: String,
    pub canonical_filter: String,
    pub order_by: String,
    pub facet_specs: Vec<FacetSpec>,
    pub dynamic_facet_spec: Option<DynamicFacetSpec>,
    pub boost_spec: Option<BoostSpec>,
    pub query_expansion_spec: Option<QueryExpansionSpec>,
    pub relevance_threshold: i32,
    pub variant_rollup_keys: Vec<String>,
    pub page_categories: Vec<String>,
    pub search_mode: i32,
    pub personalization_spec: Option<PersonalizationSpec>,
    pub labels: BTreeMap<String, String>,
    pub spell_correction_spec: Option<SpellCorrectionSpec>,
    pub entity: String,

}

Request message for [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] method.

Fields
placement: String

Required. The resource name of the Retail Search serving config, such as projects/*/locations/global/catalogs/default_catalog/servingConfigs/default_serving_config or the name of the legacy placement resource, such as projects/*/locations/global/catalogs/default_catalog/placements/default_search. This field is used to identify the serving config name and the set of models that are used to make the search.

branch: String

The branch resource name, such as projects/*/locations/global/catalogs/default_catalog/branches/0.

Use “default_branch” as the branch ID or leave this field empty, to search products under the default branch.

query: String

Raw search query.

If this field is empty, the request is considered a category browsing request and returned results are based on [filter][google.cloud.retail.v2alpha.SearchRequest.filter] and [page_categories][google.cloud.retail.v2alpha.SearchRequest.page_categories].

visitor_id: String

Required. A unique identifier for tracking visitors. For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor logs in or out of the website.

This should be the same identifier as [UserEvent.visitor_id][google.cloud.retail.v2alpha.UserEvent.visitor_id].

The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.

user_info: Option<UserInfo>

User information.

page_size: i32

Maximum number of [Product][google.cloud.retail.v2alpha.Product]s to return. If unspecified, defaults to a reasonable value. The maximum allowed value is 120. Values above 120 will be coerced to 120.

If this field is negative, an INVALID_ARGUMENT is returned.

page_token: String

A page token [SearchResponse.next_page_token][google.cloud.retail.v2alpha.SearchResponse.next_page_token], received from a previous [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] call. Provide this to retrieve the subsequent page.

When paginating, all other parameters provided to [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] must match the call that provided the page token. Otherwise, an INVALID_ARGUMENT error is returned.

offset: i32

A 0-indexed integer that specifies the current offset (that is, starting result location, amongst the [Product][google.cloud.retail.v2alpha.Product]s deemed by the API as relevant) in search results. This field is only considered if [page_token][google.cloud.retail.v2alpha.SearchRequest.page_token] is unset.

If this field is negative, an INVALID_ARGUMENT is returned.

filter: String

The filter syntax consists of an expression language for constructing a predicate from one or more fields of the products being filtered. Filter expression is case-sensitive. For more information, see Filter.

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

canonical_filter: String

The default filter that is applied when a user performs a search without checking any filters on the search page.

The filter applied to every search request when quality improvement such as query expansion is needed. In the case a query does not have a sufficient amount of results this filter will be used to determine whether or not to enable the query expansion flow. The original filter will still be used for the query expanded search. This field is strongly recommended to achieve high search quality.

For more information about filter syntax, see [SearchRequest.filter][google.cloud.retail.v2alpha.SearchRequest.filter].

order_by: String

The order in which products are returned. Products can be ordered by a field in an [Product][google.cloud.retail.v2alpha.Product] object. Leave it unset if ordered by relevance. OrderBy expression is case-sensitive. For more information, see Order.

If this field is unrecognizable, an INVALID_ARGUMENT is returned.

facet_specs: Vec<FacetSpec>

Facet specifications for faceted search. If empty, no facets are returned.

A maximum of 200 values are allowed. Otherwise, an INVALID_ARGUMENT error is returned.

dynamic_facet_spec: Option<DynamicFacetSpec>
👎Deprecated

Deprecated. Refer to https://cloud.google.com/retail/docs/configs#dynamic to enable dynamic facets. Do not set this field.

The specification for dynamically generated facets. Notice that only textual facets can be dynamically generated.

boost_spec: Option<BoostSpec>

Boost specification to boost certain products. For more information, see Boost results.

Notice that if both [ServingConfig.boost_control_ids][google.cloud.retail.v2alpha.ServingConfig.boost_control_ids] and [SearchRequest.boost_spec][google.cloud.retail.v2alpha.SearchRequest.boost_spec] are set, the boost conditions from both places are evaluated. If a search request matches multiple boost conditions, the final boost score is equal to the sum of the boost scores from all matched boost conditions.

query_expansion_spec: Option<QueryExpansionSpec>

The query expansion specification that specifies the conditions under which query expansion occurs. For more information, see Query expansion.

relevance_threshold: i32

The relevance threshold of the search results.

Defaults to [RelevanceThreshold.HIGH][google.cloud.retail.v2alpha.SearchRequest.RelevanceThreshold.HIGH], which means only the most relevant results are shown, and the least number of results are returned. For more information, see Adjust result size.

variant_rollup_keys: Vec<String>

The keys to fetch and rollup the matching [variant][google.cloud.retail.v2alpha.Product.Type.VARIANT] [Product][google.cloud.retail.v2alpha.Product]s attributes, [FulfillmentInfo][google.cloud.retail.v2alpha.FulfillmentInfo] or [LocalInventory][google.cloud.retail.v2alpha.LocalInventory]s attributes. The attributes from all the matching [variant][google.cloud.retail.v2alpha.Product.Type.VARIANT] [Product][google.cloud.retail.v2alpha.Product]s or [LocalInventory][google.cloud.retail.v2alpha.LocalInventory]s are merged and de-duplicated. Notice that rollup attributes will lead to extra query latency. Maximum number of keys is 30.

For [FulfillmentInfo][google.cloud.retail.v2alpha.FulfillmentInfo], a fulfillment type and a fulfillment ID must be provided in the format of “fulfillmentType.fulfillmentId”. E.g., in “pickupInStore.store123”, “pickupInStore” is fulfillment type and “store123” is the store ID.

Supported keys are:

colorFamilies

price

originalPrice

discount

variantId

inventory(place_id,price)

inventory(place_id,original_price)

inventory(place_id,attributes.key), where key is any key in the [Product.local_inventories.attributes][google.cloud.retail.v2alpha.LocalInventory.attributes] map.

attributes.key, where key is any key in the [Product.attributes][google.cloud.retail.v2alpha.Product.attributes] map.

pickupInStore.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “pickup-in-store”.

shipToStore.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “ship-to-store”.

sameDayDelivery.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “same-day-delivery”.

nextDayDelivery.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “next-day-delivery”.

customFulfillment1.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “custom-type-1”.

customFulfillment2.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “custom-type-2”.

customFulfillment3.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “custom-type-3”.

customFulfillment4.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “custom-type-4”.

customFulfillment5.id, where id is any [FulfillmentInfo.place_ids][google.cloud.retail.v2alpha.FulfillmentInfo.place_ids] for [FulfillmentInfo.type][google.cloud.retail.v2alpha.FulfillmentInfo.type] “custom-type-5”.

If this field is set to an invalid value other than these, an INVALID_ARGUMENT error is returned.

page_categories: Vec<String>

The categories associated with a category page. Must be set for category navigation queries to achieve good search quality. The format should be the same as [UserEvent.page_categories][google.cloud.retail.v2alpha.UserEvent.page_categories];

To represent full path of category, use ‘>’ sign to separate different hierarchies. If ‘>’ is part of the category name, replace it with other character(s).

Category pages include special pages such as sales or promotions. For instance, a special sale page may have the category hierarchy: “pageCategories” : [“Sales > 2017 Black Friday Deals”].

search_mode: i32

The search mode of the search request. If not specified, a single search request triggers both product search and faceted search.

personalization_spec: Option<PersonalizationSpec>

The specification for personalization.

Notice that if both [ServingConfig.personalization_spec][google.cloud.retail.v2alpha.ServingConfig.personalization_spec] and [SearchRequest.personalization_spec][google.cloud.retail.v2alpha.SearchRequest.personalization_spec] are set. [SearchRequest.personalization_spec][google.cloud.retail.v2alpha.SearchRequest.personalization_spec] will override [ServingConfig.personalization_spec][google.cloud.retail.v2alpha.ServingConfig.personalization_spec].

labels: BTreeMap<String, String>

The labels applied to a resource must meet the following requirements:

Each resource can have multiple labels, up to a maximum of 64.

Each label must be a key-value pair.

Keys have a minimum length of 1 character and a maximum length of 63 characters and cannot be empty. Values can be empty and have a maximum length of 63 characters.

Keys and values can contain only lowercase letters, numeric characters, underscores, and dashes. All characters must use UTF-8 encoding, and international characters are allowed.

The key portion of a label must be unique. However, you can use the same key with multiple resources.

Keys must start with a lowercase letter or international character.

For more information, see Requirements for labels in the Resource Manager documentation.

spell_correction_spec: Option<SpellCorrectionSpec>

The spell correction specification that specifies the mode under which spell correction will take effect.

entity: String

The entity for customers that may run multiple different entities, domains, sites or regions, for example, Google US, Google Ads, Waymo, google.com, youtube.com, etc. If this is set, it should be exactly matched with [UserEvent.entity][google.cloud.retail.v2alpha.UserEvent.entity] to get search results boosted by entity.

Implementations

impl SearchRequest

pub fn relevance_threshold(&self) -> RelevanceThreshold

Returns the enum value of relevance_threshold, or the default if the field is set to an invalid enum value.

pub fn set_relevance_threshold(&mut self, value: RelevanceThreshold)

Sets relevance_threshold to the provided enum value.

pub fn search_mode(&self) -> SearchMode

Returns the enum value of search_mode, or the default if the field is set to an invalid enum value.

pub fn set_search_mode(&mut self, value: SearchMode)

Sets search_mode to the provided enum value.