Struct google_api_proto::google::cloud::retail::v2alpha::SearchResponse

pub struct SearchResponse {
    pub results: Vec<SearchResult>,
    pub facets: Vec<Facet>,
    pub total_size: i32,
    pub corrected_query: String,
    pub attribution_token: String,
    pub next_page_token: String,
    pub query_expansion_info: Option<QueryExpansionInfo>,
    pub redirect_uri: String,
    pub applied_controls: Vec<String>,
    pub invalid_condition_boost_specs: Vec<ConditionBoostSpec>,
    pub experiment_info: Vec<ExperimentInfo>,
}

Response message for [SearchService.Search][google.cloud.retail.v2alpha.SearchService.Search] method.

Fields
results: Vec<SearchResult>

A list of matched items. The order represents the ranking.

facets: Vec<Facet>

Results of facets requested by user.

total_size: i32

The estimated total count of matched items irrespective of pagination. The count of [results][google.cloud.retail.v2alpha.SearchResponse.results] returned by pagination may be less than the [total_size][google.cloud.retail.v2alpha.SearchResponse.total_size] that matches.

corrected_query: String

Contains the spell corrected query, if found. If the spell correction type is AUTOMATIC, then the search results are based on corrected_query. Otherwise the original query is used for search.

attribution_token: String

A unique search token. This should be included in the [UserEvent][google.cloud.retail.v2alpha.UserEvent] logs resulting from this search, which enables accurate attribution of search model performance.

next_page_token: String

A token that can be sent as [SearchRequest.page_token][google.cloud.retail.v2alpha.SearchRequest.page_token] to retrieve the next page. If this field is omitted, there are no subsequent pages.

query_expansion_info: Option<QueryExpansionInfo>

Query expansion information for the returned results.

redirect_uri: String

The URI of a customer-defined redirect page. If redirect action is triggered, no search is performed, and only [redirect_uri][google.cloud.retail.v2alpha.SearchResponse.redirect_uri] and [attribution_token][google.cloud.retail.v2alpha.SearchResponse.attribution_token] are set in the response.

applied_controls: Vec<String>

The fully qualified resource name of applied controls.

invalid_condition_boost_specs: Vec<ConditionBoostSpec>

The invalid [SearchRequest.BoostSpec.condition_boost_specs][google.cloud.retail.v2alpha.SearchRequest.BoostSpec.condition_boost_specs] that are not applied during serving.

experiment_info: Vec<ExperimentInfo>

Metadata related to A/B testing [Experiment][google.cloud.retail.v2alpha.Experiment] associated with this response. Only exists when an experiment is triggered.