Module google_api_proto::google::cloud::run::v2

Modules

binary_authorization

Nested message and enum types in BinaryAuthorization.

condition

Nested message and enum types in Condition.

empty_dir_volume_source

Nested message and enum types in EmptyDirVolumeSource.

env_var

Nested message and enum types in EnvVar.

execution_reference

Nested message and enum types in ExecutionReference.

executions_client

Generated client implementations.

job

Nested message and enum types in Job.

jobs_client

Generated client implementations.

probe

Nested message and enum types in Probe.

revisions_client

Generated client implementations.

run_job_request

Nested message and enum types in RunJobRequest.

services_client

Generated client implementations.

task_template

Nested message and enum types in TaskTemplate.

tasks_client

Generated client implementations.

volume

Nested message and enum types in Volume.

vpc_access

Nested message and enum types in VpcAccess.

Structs

BinaryAuthorization

Settings for Binary Authorization feature.

CancelExecutionRequest

Request message for deleting an Execution.

CloudSqlInstance

Represents a set of Cloud SQL instances. Each one will be available under /cloudsql/[instance]. Visit https://cloud.google.com/sql/docs/mysql/connect-run for more information on how to connect Cloud SQL and Cloud Run.

Condition

Defines a status condition for a resource.

Container

A single application container. This specifies both the container to run, the command to run in the container and the arguments to supply to it. Note that additional arguments can be supplied by the system to the container at runtime.

ContainerPort

ContainerPort represents a network port in a single container.

CreateJobRequest

Request message for creating a Job.

CreateServiceRequest

Request message for creating a Service.

DeleteExecutionRequest

Request message for deleting an Execution.

DeleteJobRequest

Request message to delete a Job by its full name.

DeleteRevisionRequest

Request message for deleting a retired Revision. Revision lifecycle is usually managed by making changes to the parent Service. Only retired revisions can be deleted with this API.

DeleteServiceRequest

Request message to delete a Service by its full name.

EmptyDirVolumeSource

In memory (tmpfs) ephemeral storage. It is ephemeral in the sense that when the sandbox is taken down, the data is destroyed with it (it does not persist across sandbox runs).

EnvVar

EnvVar represents an environment variable present in a Container.

EnvVarSource

EnvVarSource represents a source for the value of an EnvVar.

Execution

Execution represents the configuration of a single execution. A execution an immutable resource that references a container image which is run to completion.

ExecutionReference

Reference to an Execution. Use /Executions.GetExecution with the given name to get full execution including the latest status.

ExecutionTemplate

ExecutionTemplate describes the data an execution should have when created from a template.

GcsVolumeSource

Represents a volume backed by a Cloud Storage bucket using Cloud Storage FUSE.

GetExecutionRequest

Request message for obtaining a Execution by its full name.

GetJobRequest

Request message for obtaining a Job by its full name.

GetRevisionRequest

Request message for obtaining a Revision by its full name.

GetServiceRequest

Request message for obtaining a Service by its full name.

GetTaskRequest

Request message for obtaining a Task by its full name.

GrpcAction

GRPCAction describes an action involving a GRPC port.

HttpGetAction

HTTPGetAction describes an action based on HTTP Get requests.

HttpHeader

HTTPHeader describes a custom header to be used in HTTP probes

Job

Job represents the configuration of a single job, which references a container image that is run to completion.

ListExecutionsRequest

Request message for retrieving a list of Executions.

ListExecutionsResponse

Response message containing a list of Executions.

ListJobsRequest

Request message for retrieving a list of Jobs.

ListJobsResponse

Response message containing a list of Jobs.

ListRevisionsRequest

Request message for retrieving a list of Revisions.

ListRevisionsResponse

Response message containing a list of Revisions.

ListServicesRequest

Request message for retrieving a list of Services.

ListServicesResponse

Response message containing a list of Services.

ListTasksRequest

Request message for retrieving a list of Tasks.

ListTasksResponse

Response message containing a list of Tasks.

NfsVolumeSource

Represents an NFS mount.

Probe

Probe describes a health check to be performed against a container to determine whether it is alive or ready to receive traffic.

ResourceRequirements

ResourceRequirements describes the compute resource requirements.

Revision

A Revision is an immutable snapshot of code and configuration. A Revision references a container image. Revisions are only created by updates to its parent Service.

RevisionScaling

Settings for revision-level scaling settings.

RevisionScalingStatus

Effective settings for the current revision

RevisionTemplate

RevisionTemplate describes the data a revision should have when created from a template.

RunJobRequest

Request message to create a new Execution of a Job.

SecretKeySelector

SecretEnvVarSource represents a source for the value of an EnvVar.

SecretVolumeSource

The secret’s value will be presented as the content of a file whose name is defined in the item path. If no items are defined, the name of the file is the secret.

Service

Service acts as a top-level container that manages a set of configurations and revision templates which implement a network service. Service exists to provide a singular abstraction which can be access controlled, reasoned about, and which encapsulates software lifecycle decisions such as rollout policy and team resource ownership.

ServiceScaling

Scaling settings applied at the service level rather than at the revision level.

Task

Task represents a single run of a container to completion.

TaskAttemptResult

Result of a task attempt.

TaskTemplate

TaskTemplate describes the data a task should have when created from a template.

TcpSocketAction

TCPSocketAction describes an action based on opening a socket

TrafficTarget

Holds a single traffic routing entry for the Service. Allocations can be done to a specific Revision name, or pointing to the latest Ready Revision.

TrafficTargetStatus

Represents the observed state of a single TrafficTarget entry.

UpdateJobRequest

Request message for updating a Job.

UpdateServiceRequest

Request message for updating a service.

VersionToPath

VersionToPath maps a specific version of a secret to a relative file to mount to, relative to VolumeMount’s mount_path.

Volume

Volume represents a named volume in a container.

VolumeMount

VolumeMount describes a mounting of a Volume within a container.

VpcAccess

VPC Access settings. For more information on sending traffic to a VPC network, visit https://cloud.google.com/run/docs/configuring/connecting-vpc.

Enums

EncryptionKeyRevocationAction

Specifies behavior if an encryption key used by a resource is revoked.

ExecutionEnvironment

Alternatives for execution environments.

IngressTraffic

Allowed ingress traffic for the Container.

TrafficTargetAllocationType

The type of instance allocation.