Struct google_api_proto::google::cloud::run::v2::Job

pub struct Job {

    pub name: String,
    pub uid: String,
    pub generation: i64,
    pub labels: BTreeMap<String, String>,
    pub annotations: BTreeMap<String, String>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub delete_time: Option<Timestamp>,
    pub expire_time: Option<Timestamp>,
    pub creator: String,
    pub last_modifier: String,
    pub client: String,
    pub client_version: String,
    pub launch_stage: i32,
    pub binary_authorization: Option<BinaryAuthorization>,
    pub template: Option<ExecutionTemplate>,
    pub observed_generation: i64,
    pub terminal_condition: Option<Condition>,
    pub conditions: Vec<Condition>,
    pub execution_count: i32,
    pub latest_created_execution: Option<ExecutionReference>,
    pub reconciling: bool,
    pub satisfies_pzs: bool,
    pub etag: String,

}

Job represents the configuration of a single job, which references a container image that is run to completion.

Fields
name: String

The fully qualified name of this Job.

Format: projects/{project}/locations/{location}/jobs/{job}

uid: String

Output only. Server assigned unique identifier for the Execution. The value is a UUID4 string and guaranteed to remain unchanged until the resource is deleted.

generation: i64

Output only. A number that monotonically increases every time the user modifies the desired state.

labels: BTreeMap<String, String>

Unstructured key value map that can be used to organize and categorize objects. User-provided labels are shared with Google’s billing system, so they can be used to filter, or break down billing charges by team, component, environment, state, etc. For more information, visit https://cloud.google.com/resource-manager/docs/creating-managing-labels or https://cloud.google.com/run/docs/configuring/labels.

Cloud Run API v2 does not support labels with `run.googleapis.com`, `cloud.googleapis.com`, `serving.knative.dev`, or `autoscaling.knative.dev` namespaces, and they will be rejected. All system labels in v1 now have a corresponding field in v2 Job.
annotations: BTreeMap<String, String>

Unstructured key value map that may be set by external tools to store and arbitrary metadata. They are not queryable and should be preserved when modifying objects.

Cloud Run API v2 does not support annotations with `run.googleapis.com`, `cloud.googleapis.com`, `serving.knative.dev`, or `autoscaling.knative.dev` namespaces, and they will be rejected on new resources. All system annotations in v1 now have a corresponding field in v2 Job.
This field follows Kubernetes annotations' namespacing, limits, and rules.
create_time: Option<Timestamp>

Output only. The creation time.

update_time: Option<Timestamp>

Output only. The last-modified time.

delete_time: Option<Timestamp>

Output only. The deletion time.

expire_time: Option<Timestamp>

Output only. For a deleted resource, the time after which it will be permamently deleted.

creator: String

Output only. Email address of the authenticated creator.

last_modifier: String

Output only. Email address of the last authenticated modifier.

client: String

Arbitrary identifier for the API client.

client_version: String

Arbitrary version identifier for the API client.

launch_stage: i32

The launch stage as defined by Google Cloud Platform Launch Stages. Cloud Run supports ALPHA, BETA, and GA. If no value is specified, GA is assumed. Set the launch stage to a preview stage on input to allow use of preview features in that stage. On read (or output), describes whether the resource uses preview features.

For example, if ALPHA is provided as input, but only BETA and GA-level features are used, this field will be BETA on output.
binary_authorization: Option<BinaryAuthorization>

Settings for the Binary Authorization feature.

template: Option<ExecutionTemplate>

Required. The template used to create executions for this Job.

observed_generation: i64

Output only. The generation of this Job. See comments in reconciling for additional information on reconciliation process in Cloud Run.

terminal_condition: Option<Condition>

Output only. The Condition of this Job, containing its readiness status, and detailed error information in case it did not reach the desired state.

conditions: Vec<Condition>

Output only. The Conditions of all other associated sub-resources. They contain additional diagnostics information in case the Job does not reach its desired state. See comments in reconciling for additional information on reconciliation process in Cloud Run.

execution_count: i32

Output only. Number of executions created for this job.

latest_created_execution: Option<ExecutionReference>

Output only. Name of the last created execution.

reconciling: bool

Output only. Returns true if the Job is currently being acted upon by the system to bring it into the desired state.

When a new Job is created, or an existing one is updated, Cloud Run will asynchronously perform all necessary steps to bring the Job to the desired state. This process is called reconciliation. While reconciliation is in process, observed_generation and latest_succeeded_execution, will have transient values that might mismatch the intended state: Once reconciliation is over (and this field is false), there are two possible outcomes: reconciliation succeeded and the state matches the Job, or there was an error, and reconciliation failed. This state can be found in terminal_condition.state.

If reconciliation succeeded, the following fields will match: observed_generation and generation, latest_succeeded_execution and latest_created_execution.

If reconciliation failed, observed_generation and latest_succeeded_execution will have the state of the last succeeded execution or empty for newly created Job. Additional information on the failure can be found in terminal_condition and conditions.

satisfies_pzs: bool

Output only. Reserved for future use.

etag: String

Output only. A system-generated fingerprint for this version of the resource. May be used to detect modification conflict during updates.

Implementations

impl Job

pub fn launch_stage(&self) -> LaunchStage

Returns the enum value of launch_stage, or the default if the field is set to an invalid enum value.

pub fn set_launch_stage(&mut self, value: LaunchStage)

Sets launch_stage to the provided enum value.