Struct google_api_proto::google::cloud::run::v2::Service

pub struct Service {

    pub name: String,
    pub description: String,
    pub uid: String,
    pub generation: i64,
    pub labels: BTreeMap<String, String>,
    pub annotations: BTreeMap<String, String>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub delete_time: Option<Timestamp>,
    pub expire_time: Option<Timestamp>,
    pub creator: String,
    pub last_modifier: String,
    pub client: String,
    pub client_version: String,
    pub ingress: i32,
    pub launch_stage: i32,
    pub binary_authorization: Option<BinaryAuthorization>,
    pub template: Option<RevisionTemplate>,
    pub traffic: Vec<TrafficTarget>,
    pub scaling: Option<ServiceScaling>,
    pub default_uri_disabled: bool,
    pub observed_generation: i64,
    pub terminal_condition: Option<Condition>,
    pub conditions: Vec<Condition>,
    pub latest_ready_revision: String,
    pub latest_created_revision: String,
    pub traffic_statuses: Vec<TrafficTargetStatus>,
    pub uri: String,
    pub custom_audiences: Vec<String>,
    pub satisfies_pzs: bool,
    pub reconciling: bool,
    pub etag: String,

}

Service acts as a top-level container that manages a set of configurations and revision templates which implement a network service. Service exists to provide a singular abstraction which can be access controlled, reasoned about, and which encapsulates software lifecycle decisions such as rollout policy and team resource ownership.

Fields
name: String

The fully qualified name of this Service. In CreateServiceRequest, this field is ignored, and instead composed from CreateServiceRequest.parent and CreateServiceRequest.service_id.

Format: projects/{project}/locations/{location}/services/{service_id}

description: String

User-provided description of the Service. This field currently has a 512-character limit.

uid: String

Output only. Server assigned unique identifier for the trigger. The value is a UUID4 string and guaranteed to remain unchanged until the resource is deleted.

generation: i64

Output only. A number that monotonically increases every time the user modifies the desired state. Please note that unlike v1, this is an int64 value. As with most Google APIs, its JSON representation will be a string instead of an integer.

labels: BTreeMap<String, String>

Optional. Unstructured key value map that can be used to organize and categorize objects. User-provided labels are shared with Google’s billing system, so they can be used to filter, or break down billing charges by team, component, environment, state, etc. For more information, visit https://cloud.google.com/resource-manager/docs/creating-managing-labels or https://cloud.google.com/run/docs/configuring/labels.

Cloud Run API v2 does not support labels with `run.googleapis.com`, `cloud.googleapis.com`, `serving.knative.dev`, or `autoscaling.knative.dev` namespaces, and they will be rejected. All system labels in v1 now have a corresponding field in v2 Service.
annotations: BTreeMap<String, String>

Optional. Unstructured key value map that may be set by external tools to store and arbitrary metadata. They are not queryable and should be preserved when modifying objects.

Cloud Run API v2 does not support annotations with `run.googleapis.com`, `cloud.googleapis.com`, `serving.knative.dev`, or `autoscaling.knative.dev` namespaces, and they will be rejected in new resources. All system annotations in v1 now have a corresponding field in v2 Service.
This field follows Kubernetes annotations' namespacing, limits, and rules.
create_time: Option<Timestamp>

Output only. The creation time.

update_time: Option<Timestamp>

Output only. The last-modified time.

delete_time: Option<Timestamp>

Output only. The deletion time. It is only populated as a response to a Delete request.

expire_time: Option<Timestamp>

Output only. For a deleted resource, the time after which it will be permamently deleted.

creator: String

Output only. Email address of the authenticated creator.

last_modifier: String

Output only. Email address of the last authenticated modifier.

client: String

Arbitrary identifier for the API client.

client_version: String

Arbitrary version identifier for the API client.

ingress: i32

Optional. Provides the ingress settings for this Service. On output, returns the currently observed ingress settings, or INGRESS_TRAFFIC_UNSPECIFIED if no revision is active.

launch_stage: i32

Optional. The launch stage as defined by Google Cloud Platform Launch Stages. Cloud Run supports ALPHA, BETA, and GA. If no value is specified, GA is assumed. Set the launch stage to a preview stage on input to allow use of preview features in that stage. On read (or output), describes whether the resource uses preview features.

For example, if ALPHA is provided as input, but only BETA and GA-level features are used, this field will be BETA on output.
binary_authorization: Option<BinaryAuthorization>

Optional. Settings for the Binary Authorization feature.

template: Option<RevisionTemplate>

Required. The template used to create revisions for this Service.

traffic: Vec<TrafficTarget>

Optional. Specifies how to distribute traffic over a collection of Revisions belonging to the Service. If traffic is empty or not provided, defaults to 100% traffic to the latest Ready Revision.

scaling: Option<ServiceScaling>

Optional. Specifies service-level scaling settings

default_uri_disabled: bool

Optional. Disables public resolution of the default URI of this service.

observed_generation: i64

Output only. The generation of this Service currently serving traffic. See comments in reconciling for additional information on reconciliation process in Cloud Run. Please note that unlike v1, this is an int64 value. As with most Google APIs, its JSON representation will be a string instead of an integer.

terminal_condition: Option<Condition>

Output only. The Condition of this Service, containing its readiness status, and detailed error information in case it did not reach a serving state. See comments in reconciling for additional information on reconciliation process in Cloud Run.

conditions: Vec<Condition>

Output only. The Conditions of all other associated sub-resources. They contain additional diagnostics information in case the Service does not reach its Serving state. See comments in reconciling for additional information on reconciliation process in Cloud Run.

latest_ready_revision: String

Output only. Name of the latest revision that is serving traffic. See comments in reconciling for additional information on reconciliation process in Cloud Run.

latest_created_revision: String

Output only. Name of the last created revision. See comments in reconciling for additional information on reconciliation process in Cloud Run.

traffic_statuses: Vec<TrafficTargetStatus>

Output only. Detailed status information for corresponding traffic targets. See comments in reconciling for additional information on reconciliation process in Cloud Run.

uri: String

Output only. The main URI in which this Service is serving traffic.

custom_audiences: Vec<String>

One or more custom audiences that you want this service to support. Specify each custom audience as the full URL in a string. The custom audiences are encoded in the token and used to authenticate requests. For more information, see https://cloud.google.com/run/docs/configuring/custom-audiences.

satisfies_pzs: bool

Output only. Reserved for future use.

reconciling: bool

Output only. Returns true if the Service is currently being acted upon by the system to bring it into the desired state.

When a new Service is created, or an existing one is updated, Cloud Run will asynchronously perform all necessary steps to bring the Service to the desired serving state. This process is called reconciliation. While reconciliation is in process, observed_generation, latest_ready_revison, traffic_statuses, and uri will have transient values that might mismatch the intended state: Once reconciliation is over (and this field is false), there are two possible outcomes: reconciliation succeeded and the serving state matches the Service, or there was an error, and reconciliation failed. This state can be found in terminal_condition.state.

If reconciliation succeeded, the following fields will match: traffic and traffic_statuses, observed_generation and generation, latest_ready_revision and latest_created_revision.

If reconciliation failed, traffic_statuses, observed_generation, and latest_ready_revision will have the state of the last serving revision, or empty for newly created Services. Additional information on the failure can be found in terminal_condition and conditions.

etag: String

Output only. A system-generated fingerprint for this version of the resource. May be used to detect modification conflict during updates.

Implementations

impl Service

pub fn ingress(&self) -> IngressTraffic

Returns the enum value of ingress, or the default if the field is set to an invalid enum value.

pub fn set_ingress(&mut self, value: IngressTraffic)

Sets ingress to the provided enum value.

pub fn launch_stage(&self) -> LaunchStage

Returns the enum value of launch_stage, or the default if the field is set to an invalid enum value.

pub fn set_launch_stage(&mut self, value: LaunchStage)

Sets launch_stage to the provided enum value.