Struct google_api_proto::google::cloud::runtimeconfig::v1beta1::runtime_config_manager_client::RuntimeConfigManagerClient

pub struct RuntimeConfigManagerClient<T> { /* private fields */ }

RuntimeConfig API represents configuration objects and operations on those configuration objects. RuntimeConfig objects consist of Variables logically grouped in the those objects. Variables are simple key-value pairs. Variables can be watched for changes or deletions. Variable key can be hieararchical, e.g. ports/serving_port, ports/monitoring_port, etc. Variable names can be hierarchical. No variable name can be prefix of another. Config objects represent logical containers for variables, e.g. flags, passwords, etc.

Implementations

impl<T> RuntimeConfigManagerClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> RuntimeConfigManagerClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_configs(
&mut self,
request: impl IntoRequest<ListConfigsRequest>
) -> Result<Response<ListConfigsResponse>, Status>

Lists all the RuntimeConfig resources within project.

pub async fn get_config(
&mut self,
request: impl IntoRequest<GetConfigRequest>
) -> Result<Response<RuntimeConfig>, Status>

Gets information about a RuntimeConfig resource.

pub async fn create_config(
&mut self,
request: impl IntoRequest<CreateConfigRequest>
) -> Result<Response<RuntimeConfig>, Status>

Creates a new RuntimeConfig resource. The configuration name must be unique within project.

pub async fn update_config(
&mut self,
request: impl IntoRequest<UpdateConfigRequest>
) -> Result<Response<RuntimeConfig>, Status>

Updates a RuntimeConfig resource. The configuration must exist beforehand.

pub async fn delete_config(
&mut self,
request: impl IntoRequest<DeleteConfigRequest>
) -> Result<Response<()>, Status>

Deletes a RuntimeConfig resource.

pub async fn list_variables(
&mut self,
request: impl IntoRequest<ListVariablesRequest>
) -> Result<Response<ListVariablesResponse>, Status>

Lists variables within given a configuration, matching any provided filters. This only lists variable names, not the values, unless return_values is true, in which case only variables that user has IAM permission to GetVariable will be returned.

pub async fn get_variable(
&mut self,
request: impl IntoRequest<GetVariableRequest>
) -> Result<Response<Variable>, Status>

Gets information about a single variable.

pub async fn watch_variable(
&mut self,
request: impl IntoRequest<WatchVariableRequest>
) -> Result<Response<Variable>, Status>

Watches a specific variable and waits for a change in the variable’s value. When there is a change, this method returns the new value or times out.

If a variable is deleted while being watched, the variableState state is set to DELETED and the method returns the last known variable value.

If you set the deadline for watching to a larger value than internal timeout (60 seconds), the current variable value is returned and the variableState will be VARIABLE_STATE_UNSPECIFIED.

To learn more about creating a watcher, read the Watching a Variable for Changes documentation.

pub async fn create_variable(
&mut self,
request: impl IntoRequest<CreateVariableRequest>
) -> Result<Response<Variable>, Status>

Creates a variable within the given configuration. You cannot create a variable with a name that is a prefix of an existing variable name, or a name that has an existing variable name as a prefix.

To learn more about creating a variable, read the Setting and Getting Data documentation.

pub async fn update_variable(
&mut self,
request: impl IntoRequest<UpdateVariableRequest>
) -> Result<Response<Variable>, Status>

Updates an existing variable with a new value.

pub async fn delete_variable(
&mut self,
request: impl IntoRequest<DeleteVariableRequest>
) -> Result<Response<()>, Status>

Deletes a variable or multiple variables.

If you specify a variable name, then that variable is deleted. If you specify a prefix and recursive is true, then all variables with that prefix are deleted. You must set a recursive to true if you delete variables by prefix.

pub async fn list_waiters(
&mut self,
request: impl IntoRequest<ListWaitersRequest>
) -> Result<Response<ListWaitersResponse>, Status>

List waiters within the given configuration.

pub async fn get_waiter(
&mut self,
request: impl IntoRequest<GetWaiterRequest>
) -> Result<Response<Waiter>, Status>

Gets information about a single waiter.

pub async fn create_waiter(
&mut self,
request: impl IntoRequest<CreateWaiterRequest>
) -> Result<Response<Operation>, Status>

Creates a Waiter resource. This operation returns a long-running Operation resource which can be polled for completion. However, a waiter with the given name will exist (and can be retrieved) prior to the operation completing. If the operation fails, the failed Waiter resource will still exist and must be deleted prior to subsequent creation attempts.

pub async fn delete_waiter(
&mut self,
request: impl IntoRequest<DeleteWaiterRequest>
) -> Result<Response<()>, Status>

Deletes the waiter with the specified name.