Struct google_api_proto::google::cloud::runtimeconfig::v1beta1::Variable

pub struct Variable {
    pub name: String,
    pub update_time: Option<Timestamp>,
    pub state: i32,
    pub contents: Option<Contents>,
}

Describes a single variable within a RuntimeConfig resource. The name denotes the hierarchical variable name. For example, ports/serving_port is a valid variable name. The variable value is an opaque string and only leaf variables can have values (that is, variables that do not have any child variables).

Fields
name: String

The name of the variable resource, in the format:

 projects/\[PROJECT_ID]/configs/[CONFIG_NAME]/variables/[VARIABLE_NAME\]

The \[PROJECT_ID\] must be a valid project ID, \[CONFIG_NAME\] must be a valid RuntimeConfig reource and \[VARIABLE_NAME\] follows Unix file system file path naming.

The \[VARIABLE_NAME\] can contain ASCII letters, numbers, slashes and dashes. Slashes are used as path element separators and are not part of the \[VARIABLE_NAME\] itself, so \[VARIABLE_NAME\] must contain at least one non-slash character. Multiple slashes are coalesced into single slash character. Each path segment should follow RFC 1035 segment specification. The length of a \[VARIABLE_NAME\] must be less than 256 bytes.

Once you create a variable, you cannot change the variable name.

update_time: Option<Timestamp>

[Output Only] The time of the last variable update.

state: i32

[Ouput only] The current state of the variable. The variable state indicates the outcome of the variables().watch call and is visible through the get and list calls.

contents: Option<Contents>

The value of the variable. It can be either a binary or a string value. You must specify one of either value or text. Specifying both will cause the server to return an error.

Implementations

impl Variable

pub fn state(&self) -> VariableState

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: VariableState)

Sets state to the provided enum value.