Struct google_api_proto::google::cloud::security::privateca::v1::certificate_authority_service_client::CertificateAuthorityServiceClient

pub struct CertificateAuthorityServiceClient<T> { /* private fields */ }

[Certificate Authority Service][google.cloud.security.privateca.v1.CertificateAuthorityService] manages private certificate authorities and issued certificates.

Implementations

impl<T> CertificateAuthorityServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> CertificateAuthorityServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_certificate( &mut self, request: impl IntoRequest<CreateCertificateRequest> ) -> Result<Response<Certificate>, Status>

Create a new [Certificate][google.cloud.security.privateca.v1.Certificate] in a given Project, Location from a particular [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn get_certificate( &mut self, request: impl IntoRequest<GetCertificateRequest> ) -> Result<Response<Certificate>, Status>

Returns a [Certificate][google.cloud.security.privateca.v1.Certificate].

pub async fn list_certificates( &mut self, request: impl IntoRequest<ListCertificatesRequest> ) -> Result<Response<ListCertificatesResponse>, Status>

Lists [Certificates][google.cloud.security.privateca.v1.Certificate].

pub async fn revoke_certificate( &mut self, request: impl IntoRequest<RevokeCertificateRequest> ) -> Result<Response<Certificate>, Status>

Revoke a [Certificate][google.cloud.security.privateca.v1.Certificate].

pub async fn update_certificate( &mut self, request: impl IntoRequest<UpdateCertificateRequest> ) -> Result<Response<Certificate>, Status>

Update a [Certificate][google.cloud.security.privateca.v1.Certificate]. Currently, the only field you can update is the [labels][google.cloud.security.privateca.v1.Certificate.labels] field.

pub async fn activate_certificate_authority( &mut self, request: impl IntoRequest<ActivateCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Activate a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] that is in state [AWAITING_USER_ACTIVATION][google.cloud.security.privateca.v1.CertificateAuthority.State.AWAITING_USER_ACTIVATION] and is of type [SUBORDINATE][google.cloud.security.privateca.v1.CertificateAuthority.Type.SUBORDINATE]. After the parent Certificate Authority signs a certificate signing request from [FetchCertificateAuthorityCsr][google.cloud.security.privateca.v1.CertificateAuthorityService.FetchCertificateAuthorityCsr], this method can complete the activation process.

pub async fn create_certificate_authority( &mut self, request: impl IntoRequest<CreateCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Create a new [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] in a given Project and Location.

pub async fn disable_certificate_authority( &mut self, request: impl IntoRequest<DisableCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Disable a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn enable_certificate_authority( &mut self, request: impl IntoRequest<EnableCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Enable a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn fetch_certificate_authority_csr( &mut self, request: impl IntoRequest<FetchCertificateAuthorityCsrRequest> ) -> Result<Response<FetchCertificateAuthorityCsrResponse>, Status>

Fetch a certificate signing request (CSR) from a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] that is in state [AWAITING_USER_ACTIVATION][google.cloud.security.privateca.v1.CertificateAuthority.State.AWAITING_USER_ACTIVATION] and is of type [SUBORDINATE][google.cloud.security.privateca.v1.CertificateAuthority.Type.SUBORDINATE]. The CSR must then be signed by the desired parent Certificate Authority, which could be another [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] resource, or could be an on-prem certificate authority. See also [ActivateCertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthorityService.ActivateCertificateAuthority].

pub async fn get_certificate_authority( &mut self, request: impl IntoRequest<GetCertificateAuthorityRequest> ) -> Result<Response<CertificateAuthority>, Status>

Returns a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn list_certificate_authorities( &mut self, request: impl IntoRequest<ListCertificateAuthoritiesRequest> ) -> Result<Response<ListCertificateAuthoritiesResponse>, Status>

Lists [CertificateAuthorities][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn undelete_certificate_authority( &mut self, request: impl IntoRequest<UndeleteCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Undelete a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] that has been deleted.

pub async fn delete_certificate_authority( &mut self, request: impl IntoRequest<DeleteCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Delete a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn update_certificate_authority( &mut self, request: impl IntoRequest<UpdateCertificateAuthorityRequest> ) -> Result<Response<Operation>, Status>

Update a [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority].

pub async fn create_ca_pool( &mut self, request: impl IntoRequest<CreateCaPoolRequest> ) -> Result<Response<Operation>, Status>

Create a [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn update_ca_pool( &mut self, request: impl IntoRequest<UpdateCaPoolRequest> ) -> Result<Response<Operation>, Status>

Update a [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn get_ca_pool( &mut self, request: impl IntoRequest<GetCaPoolRequest> ) -> Result<Response<CaPool>, Status>

Returns a [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn list_ca_pools( &mut self, request: impl IntoRequest<ListCaPoolsRequest> ) -> Result<Response<ListCaPoolsResponse>, Status>

Lists [CaPools][google.cloud.security.privateca.v1.CaPool].

pub async fn delete_ca_pool( &mut self, request: impl IntoRequest<DeleteCaPoolRequest> ) -> Result<Response<Operation>, Status>

Delete a [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn fetch_ca_certs( &mut self, request: impl IntoRequest<FetchCaCertsRequest> ) -> Result<Response<FetchCaCertsResponse>, Status>

FetchCaCerts returns the current trust anchor for the [CaPool][google.cloud.security.privateca.v1.CaPool]. This will include CA certificate chains for all ACTIVE [CertificateAuthority][google.cloud.security.privateca.v1.CertificateAuthority] resources in the [CaPool][google.cloud.security.privateca.v1.CaPool].

pub async fn get_certificate_revocation_list( &mut self, request: impl IntoRequest<GetCertificateRevocationListRequest> ) -> Result<Response<CertificateRevocationList>, Status>

Returns a [CertificateRevocationList][google.cloud.security.privateca.v1.CertificateRevocationList].

pub async fn list_certificate_revocation_lists( &mut self, request: impl IntoRequest<ListCertificateRevocationListsRequest> ) -> Result<Response<ListCertificateRevocationListsResponse>, Status>

Lists [CertificateRevocationLists][google.cloud.security.privateca.v1.CertificateRevocationList].

pub async fn update_certificate_revocation_list( &mut self, request: impl IntoRequest<UpdateCertificateRevocationListRequest> ) -> Result<Response<Operation>, Status>

Update a [CertificateRevocationList][google.cloud.security.privateca.v1.CertificateRevocationList].

pub async fn create_certificate_template( &mut self, request: impl IntoRequest<CreateCertificateTemplateRequest> ) -> Result<Response<Operation>, Status>

Create a new [CertificateTemplate][google.cloud.security.privateca.v1.CertificateTemplate] in a given Project and Location.

pub async fn delete_certificate_template( &mut self, request: impl IntoRequest<DeleteCertificateTemplateRequest> ) -> Result<Response<Operation>, Status>

DeleteCertificateTemplate deletes a [CertificateTemplate][google.cloud.security.privateca.v1.CertificateTemplate].

pub async fn get_certificate_template( &mut self, request: impl IntoRequest<GetCertificateTemplateRequest> ) -> Result<Response<CertificateTemplate>, Status>

Returns a [CertificateTemplate][google.cloud.security.privateca.v1.CertificateTemplate].

pub async fn list_certificate_templates( &mut self, request: impl IntoRequest<ListCertificateTemplatesRequest> ) -> Result<Response<ListCertificateTemplatesResponse>, Status>

Lists [CertificateTemplates][google.cloud.security.privateca.v1.CertificateTemplate].

pub async fn update_certificate_template( &mut self, request: impl IntoRequest<UpdateCertificateTemplateRequest> ) -> Result<Response<Operation>, Status>

Update a [CertificateTemplate][google.cloud.security.privateca.v1.CertificateTemplate].