Struct google_api_proto::google::cloud::security::privateca::v1beta1::certificate_authority_service_client::CertificateAuthorityServiceClient

pub struct CertificateAuthorityServiceClient<T> { /* private fields */ }

[Certificate Authority Service][google.cloud.security.privateca.v1beta1.CertificateAuthorityService] manages private certificate authorities and issued certificates.

Implementations

impl<T> CertificateAuthorityServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> CertificateAuthorityServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_certificate(
&mut self,
request: impl IntoRequest<CreateCertificateRequest>
) -> Result<Response<Certificate>, Status>

Create a new [Certificate][google.cloud.security.privateca.v1beta1.Certificate] in a given Project, Location from a particular [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn get_certificate(
&mut self,
request: impl IntoRequest<GetCertificateRequest>
) -> Result<Response<Certificate>, Status>

Returns a [Certificate][google.cloud.security.privateca.v1beta1.Certificate].

pub async fn list_certificates(
&mut self,
request: impl IntoRequest<ListCertificatesRequest>
) -> Result<Response<ListCertificatesResponse>, Status>

Lists [Certificates][google.cloud.security.privateca.v1beta1.Certificate].

pub async fn revoke_certificate(
&mut self,
request: impl IntoRequest<RevokeCertificateRequest>
) -> Result<Response<Certificate>, Status>

Revoke a [Certificate][google.cloud.security.privateca.v1beta1.Certificate].

pub async fn update_certificate(
&mut self,
request: impl IntoRequest<UpdateCertificateRequest>
) -> Result<Response<Certificate>, Status>

Update a [Certificate][google.cloud.security.privateca.v1beta1.Certificate]. Currently, the only field you can update is the [labels][google.cloud.security.privateca.v1beta1.Certificate.labels] field.

pub async fn activate_certificate_authority(
&mut self,
request: impl IntoRequest<ActivateCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Activate a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] that is in state [PENDING_ACTIVATION][google.cloud.security.privateca.v1beta1.CertificateAuthority.State.PENDING_ACTIVATION] and is of type [SUBORDINATE][google.cloud.security.privateca.v1beta1.CertificateAuthority.Type.SUBORDINATE]. After the parent Certificate Authority signs a certificate signing request from [FetchCertificateAuthorityCsr][google.cloud.security.privateca.v1beta1.CertificateAuthorityService.FetchCertificateAuthorityCsr], this method can complete the activation process.

pub async fn create_certificate_authority(
&mut self,
request: impl IntoRequest<CreateCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Create a new [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] in a given Project and Location.

pub async fn disable_certificate_authority(
&mut self,
request: impl IntoRequest<DisableCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Disable a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn enable_certificate_authority(
&mut self,
request: impl IntoRequest<EnableCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Enable a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn fetch_certificate_authority_csr(
&mut self,
request: impl IntoRequest<FetchCertificateAuthorityCsrRequest>
) -> Result<Response<FetchCertificateAuthorityCsrResponse>, Status>

Fetch a certificate signing request (CSR) from a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] that is in state [PENDING_ACTIVATION][google.cloud.security.privateca.v1beta1.CertificateAuthority.State.PENDING_ACTIVATION] and is of type [SUBORDINATE][google.cloud.security.privateca.v1beta1.CertificateAuthority.Type.SUBORDINATE]. The CSR must then be signed by the desired parent Certificate Authority, which could be another [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] resource, or could be an on-prem certificate authority. See also [ActivateCertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthorityService.ActivateCertificateAuthority].

pub async fn get_certificate_authority(
&mut self,
request: impl IntoRequest<GetCertificateAuthorityRequest>
) -> Result<Response<CertificateAuthority>, Status>

Returns a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn list_certificate_authorities(
&mut self,
request: impl IntoRequest<ListCertificateAuthoritiesRequest>
) -> Result<Response<ListCertificateAuthoritiesResponse>, Status>

Lists [CertificateAuthorities][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn restore_certificate_authority(
&mut self,
request: impl IntoRequest<RestoreCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Restore a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] that is scheduled for deletion.

pub async fn schedule_delete_certificate_authority(
&mut self,
request: impl IntoRequest<ScheduleDeleteCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Schedule a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] for deletion.

pub async fn update_certificate_authority(
&mut self,
request: impl IntoRequest<UpdateCertificateAuthorityRequest>
) -> Result<Response<Operation>, Status>

Update a [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pub async fn get_certificate_revocation_list(
&mut self,
request: impl IntoRequest<GetCertificateRevocationListRequest>
) -> Result<Response<CertificateRevocationList>, Status>

Returns a [CertificateRevocationList][google.cloud.security.privateca.v1beta1.CertificateRevocationList].

pub async fn list_certificate_revocation_lists(
&mut self,
request: impl IntoRequest<ListCertificateRevocationListsRequest>
) -> Result<Response<ListCertificateRevocationListsResponse>, Status>

Lists [CertificateRevocationLists][google.cloud.security.privateca.v1beta1.CertificateRevocationList].

pub async fn update_certificate_revocation_list(
&mut self,
request: impl IntoRequest<UpdateCertificateRevocationListRequest>
) -> Result<Response<Operation>, Status>

Update a [CertificateRevocationList][google.cloud.security.privateca.v1beta1.CertificateRevocationList].

pub async fn get_reusable_config(
&mut self,
request: impl IntoRequest<GetReusableConfigRequest>
) -> Result<Response<ReusableConfig>, Status>

Returns a [ReusableConfig][google.cloud.security.privateca.v1beta1.ReusableConfig].

pub async fn list_reusable_configs(
&mut self,
request: impl IntoRequest<ListReusableConfigsRequest>
) -> Result<Response<ListReusableConfigsResponse>, Status>

Lists [ReusableConfigs][google.cloud.security.privateca.v1beta1.ReusableConfig].