Struct google_api_proto::google::cloud::security::privateca::v1beta1::CertificateAuthority

pub struct CertificateAuthority {

    pub name: String,
    pub type: i32,
    pub tier: i32,
    pub config: Option<CertificateConfig>,
    pub lifetime: Option<Duration>,
    pub key_spec: Option<KeyVersionSpec>,
    pub certificate_policy: Option<CertificateAuthorityPolicy>,
    pub issuing_options: Option<IssuingOptions>,
    pub subordinate_config: Option<SubordinateConfig>,
    pub state: i32,
    pub pem_ca_certificates: Vec<String>,
    pub ca_certificate_descriptions: Vec<CertificateDescription>,
    pub gcs_bucket: String,
    pub access_urls: Option<AccessUrls>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub delete_time: Option<Timestamp>,
    pub labels: BTreeMap<String, String>,

}

A [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] represents an individual Certificate Authority. A [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] can be used to create [Certificates][google.cloud.security.privateca.v1beta1.Certificate].

Fields
name: String

Output only. The resource name for this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] in the format projects/*/locations/*/certificateAuthorities/*.

type: i32

Required. Immutable. The [Type][google.cloud.security.privateca.v1beta1.CertificateAuthority.Type] of this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

tier: i32

Required. Immutable. The [Tier][google.cloud.security.privateca.v1beta1.CertificateAuthority.Tier] of this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

config: Option<CertificateConfig>

Required. Immutable. The config used to create a self-signed X.509 certificate or CSR.

lifetime: Option<Duration>

Required. The desired lifetime of the CA certificate. Used to create the “not_before_time” and “not_after_time” fields inside an X.509 certificate.

key_spec: Option<KeyVersionSpec>

Required. Immutable. Used when issuing certificates for this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority]. If this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] is a self-signed CertificateAuthority, this key is also used to sign the self-signed CA certificate. Otherwise, it is used to sign a CSR.

certificate_policy: Option<CertificateAuthorityPolicy>

Optional. The [CertificateAuthorityPolicy][google.cloud.security.privateca.v1beta1.CertificateAuthority.CertificateAuthorityPolicy] to enforce when issuing [Certificates][google.cloud.security.privateca.v1beta1.Certificate] from this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

issuing_options: Option<IssuingOptions>

Optional. The [IssuingOptions][google.cloud.security.privateca.v1beta1.CertificateAuthority.IssuingOptions] to follow when issuing [Certificates][google.cloud.security.privateca.v1beta1.Certificate] from this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

subordinate_config: Option<SubordinateConfig>

Optional. If this is a subordinate [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority], this field will be set with the subordinate configuration, which describes its issuers. This may be updated, but this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] must continue to validate.

state: i32

Output only. The [State][google.cloud.security.privateca.v1beta1.CertificateAuthority.State] for this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority].

pem_ca_certificates: Vec<String>

Output only. This [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority]’s certificate chain, including the current [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority]’s certificate. Ordered such that the root issuer is the final element (consistent with RFC 5246). For a self-signed CA, this will only list the current [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority]’s certificate.

ca_certificate_descriptions: Vec<CertificateDescription>

Output only. A structured description of this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority]’s CA certificate and its issuers. Ordered as self-to-root.

gcs_bucket: String

Immutable. The name of a Cloud Storage bucket where this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] will publish content, such as the CA certificate and CRLs. This must be a bucket name, without any prefixes (such as gs://) or suffixes (such as .googleapis.com). For example, to use a bucket named my-bucket, you would simply specify my-bucket. If not specified, a managed bucket will be created.

access_urls: Option<AccessUrls>

Output only. URLs for accessing content published by this CA, such as the CA certificate and CRLs.

create_time: Option<Timestamp>

Output only. The time at which this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] was created.

update_time: Option<Timestamp>

Output only. The time at which this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] was updated.

delete_time: Option<Timestamp>

Output only. The time at which this [CertificateAuthority][google.cloud.security.privateca.v1beta1.CertificateAuthority] will be deleted, if scheduled for deletion.

labels: BTreeMap<String, String>

Optional. Labels with user-defined metadata.

Implementations

impl CertificateAuthority

pub fn type(&self) -> Type

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: Type)

Sets type to the provided enum value.

pub fn tier(&self) -> Tier

Returns the enum value of tier, or the default if the field is set to an invalid enum value.

pub fn set_tier(&mut self, value: Tier)

Sets tier to the provided enum value.

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.