Struct google_api_proto::google::cloud::securitycenter::settings::v1beta1::security_center_settings_service_client::SecurityCenterSettingsServiceClient

pub struct SecurityCenterSettingsServiceClient<T> { /* private fields */ }

API Overview

The SecurityCenterSettingsService is a sub-api of securitycenter.googleapis.com. The service provides methods to manage Security Center Settings, and Component Settings for GCP organizations, folders, projects, and clusters.

Implementations

impl<T> SecurityCenterSettingsServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> SecurityCenterSettingsServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_service_account( &mut self, request: impl IntoRequest<GetServiceAccountRequest> ) -> Result<Response<ServiceAccount>, Status>

Retrieves the organizations service account, if it exists, otherwise it creates the organization service account. This API is idempotent and will only create a service account once. On subsequent calls it will return the previously created service account. SHA, SCC and CTD Infra Automation will use this SA. This SA will not have any permissions when created. The UI will provision this via IAM or the user will using their own internal process. This API only creates SAs on the organization. Folders are not supported and projects will use per-project SAs associated with APIs enabled on a project. This API will be called by the UX onboarding workflow.

pub async fn get_settings( &mut self, request: impl IntoRequest<GetSettingsRequest> ) -> Result<Response<Settings>, Status>

Gets the Settings.

pub async fn update_settings( &mut self, request: impl IntoRequest<UpdateSettingsRequest> ) -> Result<Response<Settings>, Status>

Updates the Settings.

pub async fn reset_settings( &mut self, request: impl IntoRequest<ResetSettingsRequest> ) -> Result<Response<()>, Status>

Reset the organization, folder or project’s settings and return the settings of just that resource to the default.

Settings are present at the organization, folder, project, and cluster levels. Using Reset on a sub-organization level will remove that resource’s override and result in the parent’s settings being used (eg: if Reset on a cluster, project settings will be used).

Using Reset on organization will remove the override that was set and result in default settings being used.

pub async fn batch_get_settings( &mut self, request: impl IntoRequest<BatchGetSettingsRequest> ) -> Result<Response<BatchGetSettingsResponse>, Status>

Gets a list of settings.

pub async fn calculate_effective_settings( &mut self, request: impl IntoRequest<CalculateEffectiveSettingsRequest> ) -> Result<Response<Settings>, Status>

CalculateEffectiveSettings looks up all of the Security Center Settings resources in the GCP resource hierarchy, and calculates the effective settings on that resource by applying the following rules:

Settings provided closer to the target resource take precedence over those further away (e.g. folder will override organization level settings).

Product defaults can be overridden at org, folder, project, and cluster levels.

Detectors will be filtered out if they belong to a billing tier the customer has not configured.

pub async fn batch_calculate_effective_settings( &mut self, request: impl IntoRequest<BatchCalculateEffectiveSettingsRequest> ) -> Result<Response<BatchCalculateEffectiveSettingsResponse>, Status>

Gets a list of effective settings.

pub async fn get_component_settings( &mut self, request: impl IntoRequest<GetComponentSettingsRequest> ) -> Result<Response<ComponentSettings>, Status>

Gets the Component Settings.

pub async fn update_component_settings( &mut self, request: impl IntoRequest<UpdateComponentSettingsRequest> ) -> Result<Response<ComponentSettings>, Status>

Updates the Component Settings.

pub async fn reset_component_settings( &mut self, request: impl IntoRequest<ResetComponentSettingsRequest> ) -> Result<Response<()>, Status>

Reset the organization, folder or project’s component settings and return the settings to the default. Settings are present at the organization, folder and project levels. Using Reset for a folder or project will remove the override that was set and result in the organization-level settings being used.

pub async fn calculate_effective_component_settings( &mut self, request: impl IntoRequest<CalculateEffectiveComponentSettingsRequest> ) -> Result<Response<ComponentSettings>, Status>

Gets the Effective Component Settings.

pub async fn list_detectors( &mut self, request: impl IntoRequest<ListDetectorsRequest> ) -> Result<Response<ListDetectorsResponse>, Status>

Retrieves an unordered list of available detectors.

pub async fn list_components( &mut self, request: impl IntoRequest<ListComponentsRequest> ) -> Result<Response<ListComponentsResponse>, Status>

Retrieves an unordered list of available SCC components.