Module google_api_proto::google::cloud::securitycenter::v1::cvssv3

Nested message and enum types in Cvssv3.

Enums

AttackComplexity

This metric describes the conditions beyond the attacker’s control that must exist in order to exploit the vulnerability.

AttackVector

This metric reflects the context by which vulnerability exploitation is possible.

Impact

The Impact metrics capture the effects of a successfully exploited vulnerability on the component that suffers the worst outcome that is most directly and predictably associated with the attack.

PrivilegesRequired

This metric describes the level of privileges an attacker must possess before successfully exploiting the vulnerability.

Scope

The Scope metric captures whether a vulnerability in one vulnerable component impacts resources in components beyond its security scope.

UserInteraction

This metric captures the requirement for a human user, other than the attacker, to participate in the successful compromise of the vulnerable component.