Struct google_api_proto::google::cloud::securitycenter::v1::security_center_client::SecurityCenterClient

pub struct SecurityCenterClient<T> { /* private fields */ }

V1 APIs for Security Center service.

Implementations

impl<T> SecurityCenterClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> SecurityCenterClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn bulk_mute_findings( &mut self, request: impl IntoRequest<BulkMuteFindingsRequest>, ) -> Result<Response<Operation>, Status>

Kicks off an LRO to bulk mute findings for a parent based on a filter. The parent can be either an organization, folder or project. The findings matched by the filter will be muted after the LRO is done.

pub async fn create_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<CreateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Creates a resident SecurityHealthAnalyticsCustomModule at the scope of the given CRM parent, and also creates inherited SecurityHealthAnalyticsCustomModules for all CRM descendants of the given parent. These modules are enabled by default.

pub async fn create_source( &mut self, request: impl IntoRequest<CreateSourceRequest>, ) -> Result<Response<Source>, Status>

Creates a source.

pub async fn create_finding( &mut self, request: impl IntoRequest<CreateFindingRequest>, ) -> Result<Response<Finding>, Status>

Creates a finding. The corresponding source must exist for finding creation to succeed.

pub async fn create_mute_config( &mut self, request: impl IntoRequest<CreateMuteConfigRequest>, ) -> Result<Response<MuteConfig>, Status>

Creates a mute config.

pub async fn create_notification_config( &mut self, request: impl IntoRequest<CreateNotificationConfigRequest>, ) -> Result<Response<NotificationConfig>, Status>

Creates a notification config.

pub async fn delete_mute_config( &mut self, request: impl IntoRequest<DeleteMuteConfigRequest>, ) -> Result<Response<()>, Status>

Deletes an existing mute config.

pub async fn delete_notification_config( &mut self, request: impl IntoRequest<DeleteNotificationConfigRequest>, ) -> Result<Response<()>, Status>

Deletes a notification config.

pub async fn delete_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<DeleteSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<()>, Status>

Deletes the specified SecurityHealthAnalyticsCustomModule and all of its descendants in the CRM hierarchy. This method is only supported for resident custom modules.

pub async fn get_simulation( &mut self, request: impl IntoRequest<GetSimulationRequest>, ) -> Result<Response<Simulation>, Status>

Get the simulation by name or the latest simulation for the given organization.

pub async fn get_valued_resource( &mut self, request: impl IntoRequest<GetValuedResourceRequest>, ) -> Result<Response<ValuedResource>, Status>

Get the valued resource by name

pub async fn get_big_query_export( &mut self, request: impl IntoRequest<GetBigQueryExportRequest>, ) -> Result<Response<BigQueryExport>, Status>

Gets a BigQuery export.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Gets the access control policy on the specified Source.

pub async fn get_mute_config( &mut self, request: impl IntoRequest<GetMuteConfigRequest>, ) -> Result<Response<MuteConfig>, Status>

Gets a mute config.

pub async fn get_notification_config( &mut self, request: impl IntoRequest<GetNotificationConfigRequest>, ) -> Result<Response<NotificationConfig>, Status>

Gets a notification config.

pub async fn get_organization_settings( &mut self, request: impl IntoRequest<GetOrganizationSettingsRequest>, ) -> Result<Response<OrganizationSettings>, Status>

Gets the settings for an organization.

pub async fn get_effective_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<GetEffectiveSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<EffectiveSecurityHealthAnalyticsCustomModule>, Status>

Retrieves an EffectiveSecurityHealthAnalyticsCustomModule.

pub async fn get_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<GetSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Retrieves a SecurityHealthAnalyticsCustomModule.

pub async fn get_source( &mut self, request: impl IntoRequest<GetSourceRequest>, ) -> Result<Response<Source>, Status>

Gets a source.

pub async fn group_assets( &mut self, request: impl IntoRequest<GroupAssetsRequest>, ) -> Result<Response<GroupAssetsResponse>, Status>

Filters an organization’s assets and groups them by their specified properties.

pub async fn group_findings( &mut self, request: impl IntoRequest<GroupFindingsRequest>, ) -> Result<Response<GroupFindingsResponse>, Status>

Filters an organization or source’s findings and groups them by their specified properties.

To group across all sources provide a - as the source id. Example: /v1/organizations/{organization_id}/sources/-/findings, /v1/folders/{folder_id}/sources/-/findings, /v1/projects/{project_id}/sources/-/findings

pub async fn list_assets( &mut self, request: impl IntoRequest<ListAssetsRequest>, ) -> Result<Response<ListAssetsResponse>, Status>

Lists an organization’s assets.

pub async fn list_descendant_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListDescendantSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListDescendantSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all resident SecurityHealthAnalyticsCustomModules under the given CRM parent and all of the parent’s CRM descendants.

pub async fn list_findings( &mut self, request: impl IntoRequest<ListFindingsRequest>, ) -> Result<Response<ListFindingsResponse>, Status>

Lists an organization or source’s findings.

To list across all sources provide a - as the source id. Example: /v1/organizations/{organization_id}/sources/-/findings

pub async fn list_mute_configs( &mut self, request: impl IntoRequest<ListMuteConfigsRequest>, ) -> Result<Response<ListMuteConfigsResponse>, Status>

Lists mute configs.

pub async fn list_notification_configs( &mut self, request: impl IntoRequest<ListNotificationConfigsRequest>, ) -> Result<Response<ListNotificationConfigsResponse>, Status>

Lists notification configs.

pub async fn list_effective_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListEffectiveSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListEffectiveSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all EffectiveSecurityHealthAnalyticsCustomModules for the given parent. This includes resident modules defined at the scope of the parent, and inherited modules, inherited from CRM ancestors.

pub async fn list_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all SecurityHealthAnalyticsCustomModules for the given parent. This includes resident modules defined at the scope of the parent, and inherited modules, inherited from CRM ancestors.

pub async fn list_sources( &mut self, request: impl IntoRequest<ListSourcesRequest>, ) -> Result<Response<ListSourcesResponse>, Status>

Lists all sources belonging to an organization.

pub async fn run_asset_discovery( &mut self, request: impl IntoRequest<RunAssetDiscoveryRequest>, ) -> Result<Response<Operation>, Status>

Runs asset discovery. The discovery is tracked with a long-running operation.

This API can only be called with limited frequency for an organization. If it is called too frequently the caller will receive a TOO_MANY_REQUESTS error.

pub async fn set_finding_state( &mut self, request: impl IntoRequest<SetFindingStateRequest>, ) -> Result<Response<Finding>, Status>

Updates the state of a finding.

pub async fn set_mute( &mut self, request: impl IntoRequest<SetMuteRequest>, ) -> Result<Response<Finding>, Status>

Updates the mute state of a finding.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest>, ) -> Result<Response<Policy>, Status>

Sets the access control policy on the specified Source.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest>, ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the permissions that a caller has on the specified source.

pub async fn simulate_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<SimulateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SimulateSecurityHealthAnalyticsCustomModuleResponse>, Status>

Simulates a given SecurityHealthAnalyticsCustomModule and Resource.

pub async fn update_external_system( &mut self, request: impl IntoRequest<UpdateExternalSystemRequest>, ) -> Result<Response<ExternalSystem>, Status>

Updates external system. This is for a given finding.

pub async fn update_finding( &mut self, request: impl IntoRequest<UpdateFindingRequest>, ) -> Result<Response<Finding>, Status>

Creates or updates a finding. The corresponding source must exist for a finding creation to succeed.

pub async fn update_mute_config( &mut self, request: impl IntoRequest<UpdateMuteConfigRequest>, ) -> Result<Response<MuteConfig>, Status>

Updates a mute config.

pub async fn update_notification_config( &mut self, request: impl IntoRequest<UpdateNotificationConfigRequest>, ) -> Result<Response<NotificationConfig>, Status>

Updates a notification config. The following update fields are allowed: description, pubsub_topic, streaming_config.filter

pub async fn update_organization_settings( &mut self, request: impl IntoRequest<UpdateOrganizationSettingsRequest>, ) -> Result<Response<OrganizationSettings>, Status>

Updates an organization’s settings.

pub async fn update_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<UpdateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Updates the SecurityHealthAnalyticsCustomModule under the given name based on the given update mask. Updating the enablement state is supported on both resident and inherited modules (though resident modules cannot have an enablement state of “inherited”). Updating the display name and custom config of a module is supported on resident modules only.

pub async fn update_source( &mut self, request: impl IntoRequest<UpdateSourceRequest>, ) -> Result<Response<Source>, Status>

Updates a source.

pub async fn update_security_marks( &mut self, request: impl IntoRequest<UpdateSecurityMarksRequest>, ) -> Result<Response<SecurityMarks>, Status>

Updates security marks.

pub async fn create_big_query_export( &mut self, request: impl IntoRequest<CreateBigQueryExportRequest>, ) -> Result<Response<BigQueryExport>, Status>

Creates a BigQuery export.

pub async fn delete_big_query_export( &mut self, request: impl IntoRequest<DeleteBigQueryExportRequest>, ) -> Result<Response<()>, Status>

Deletes an existing BigQuery export.

pub async fn update_big_query_export( &mut self, request: impl IntoRequest<UpdateBigQueryExportRequest>, ) -> Result<Response<BigQueryExport>, Status>

Updates a BigQuery export.

pub async fn list_big_query_exports( &mut self, request: impl IntoRequest<ListBigQueryExportsRequest>, ) -> Result<Response<ListBigQueryExportsResponse>, Status>

Lists BigQuery exports. Note that when requesting BigQuery exports at a given level all exports under that level are also returned e.g. if requesting BigQuery exports under a folder, then all BigQuery exports immediately under the folder plus the ones created under the projects within the folder are returned.

pub async fn create_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<CreateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Creates a resident Event Threat Detection custom module at the scope of the given Resource Manager parent, and also creates inherited custom modules for all descendants of the given parent. These modules are enabled by default.

pub async fn delete_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<DeleteEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<()>, Status>

Deletes the specified Event Threat Detection custom module and all of its descendants in the Resource Manager hierarchy. This method is only supported for resident custom modules.

pub async fn get_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<GetEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Gets an Event Threat Detection custom module.

pub async fn list_descendant_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListDescendantEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListDescendantEventThreatDetectionCustomModulesResponse>, Status>

Lists all resident Event Threat Detection custom modules under the given Resource Manager parent and its descendants.

pub async fn list_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListEventThreatDetectionCustomModulesResponse>, Status>

Lists all Event Threat Detection custom modules for the given Resource Manager parent. This includes resident modules defined at the scope of the parent along with modules inherited from ancestors.

pub async fn update_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<UpdateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Updates the Event Threat Detection custom module with the given name based on the given update mask. Updating the enablement state is supported for both resident and inherited modules (though resident modules cannot have an enablement state of “inherited”). Updating the display name or configuration of a module is supported for resident modules only. The type of a module cannot be changed.

pub async fn validate_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<ValidateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<ValidateEventThreatDetectionCustomModuleResponse>, Status>

Validates the given Event Threat Detection custom module.

pub async fn get_effective_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<GetEffectiveEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EffectiveEventThreatDetectionCustomModule>, Status>

Gets an effective Event Threat Detection custom module at the given level.

pub async fn list_effective_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListEffectiveEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListEffectiveEventThreatDetectionCustomModulesResponse>, Status>

Lists all effective Event Threat Detection custom modules for the given parent. This includes resident modules defined at the scope of the parent along with modules inherited from its ancestors.

pub async fn batch_create_resource_value_configs( &mut self, request: impl IntoRequest<BatchCreateResourceValueConfigsRequest>, ) -> Result<Response<BatchCreateResourceValueConfigsResponse>, Status>

Creates a ResourceValueConfig for an organization. Maps user’s tags to difference resource values for use by the attack path simulation.

pub async fn delete_resource_value_config( &mut self, request: impl IntoRequest<DeleteResourceValueConfigRequest>, ) -> Result<Response<()>, Status>

Deletes a ResourceValueConfig.

pub async fn get_resource_value_config( &mut self, request: impl IntoRequest<GetResourceValueConfigRequest>, ) -> Result<Response<ResourceValueConfig>, Status>

Gets a ResourceValueConfig.

pub async fn list_resource_value_configs( &mut self, request: impl IntoRequest<ListResourceValueConfigsRequest>, ) -> Result<Response<ListResourceValueConfigsResponse>, Status>

Lists all ResourceValueConfigs.

pub async fn update_resource_value_config( &mut self, request: impl IntoRequest<UpdateResourceValueConfigRequest>, ) -> Result<Response<ResourceValueConfig>, Status>

Updates an existing ResourceValueConfigs with new rules.

pub async fn list_valued_resources( &mut self, request: impl IntoRequest<ListValuedResourcesRequest>, ) -> Result<Response<ListValuedResourcesResponse>, Status>

Lists the valued resources for a set of simulation results and filter.

pub async fn list_attack_paths( &mut self, request: impl IntoRequest<ListAttackPathsRequest>, ) -> Result<Response<ListAttackPathsResponse>, Status>

Lists the attack paths for a set of simulation results or valued resources and filter.