Module google_api_proto::google::cloud::securitycenter::v1p1beta1

Modules

asset

Nested message and enum types in Asset.

finding

Nested message and enum types in Finding.

list_assets_response

Nested message and enum types in ListAssetsResponse.

list_findings_response

Nested message and enum types in ListFindingsResponse.

notification_config

Nested message and enum types in NotificationConfig.

notification_message

Nested message and enum types in NotificationMessage.

organization_settings

Nested message and enum types in OrganizationSettings.

run_asset_discovery_response

Nested message and enum types in RunAssetDiscoveryResponse.

security_center_client

Generated client implementations.

Structs

Asset

Security Command Center representation of a Google Cloud resource.

CreateFindingRequest

Request message for creating a finding.

CreateNotificationConfigRequest

Request message for creating a notification config.

CreateSourceRequest

Request message for creating a source.

DeleteNotificationConfigRequest

Request message for deleting a notification config.

Finding

Security Command Center finding.

Folder

Message that contains the resource name and display name of a folder resource.

GetNotificationConfigRequest

Request message for getting a notification config.

GetOrganizationSettingsRequest

Request message for getting organization settings.

GetSourceRequest

Request message for getting a source.

GroupAssetsRequest

Request message for grouping by assets.

GroupAssetsResponse

Response message for grouping by assets.

GroupFindingsRequest

Request message for grouping by findings.

GroupFindingsResponse

Response message for group by findings.

GroupResult

Result containing the properties and count of a groupBy request.

ListAssetsRequest

Request message for listing assets.

ListAssetsResponse

Response message for listing assets.

ListFindingsRequest

Request message for listing findings.

ListFindingsResponse

Response message for listing findings.

ListNotificationConfigsRequest

Request message for listing notification configs.

ListNotificationConfigsResponse

Response message for listing notification configs.

ListSourcesRequest

Request message for listing sources.

ListSourcesResponse

Response message for listing sources.

NotificationConfig

Security Command Center notification configs.

NotificationMessage

Security Command Center’s Notification

OrganizationSettings

User specified settings that are attached to the Security Command Center organization.

Resource

Information related to the Google Cloud resource.

RunAssetDiscoveryRequest

Request message for running asset discovery for an organization.

RunAssetDiscoveryResponse

Response of asset discovery run

SecurityMarks

User specified security marks that are attached to the parent Security Command Center resource. Security marks are scoped within a Security Command Center organization – they can be modified and viewed by all users who have proper permissions on the organization.

SetFindingStateRequest

Request message for updating a finding’s state.

Source

Security Command Center finding source. A finding source is an entity or a mechanism that can produce a finding. A source is like a container of findings that come from the same scanner, logger, monitor, etc.

UpdateFindingRequest

Request message for updating or creating a finding.

UpdateNotificationConfigRequest

Request message for updating a notification config.

UpdateOrganizationSettingsRequest

Request message for updating an organization’s settings.

UpdateSecurityMarksRequest

Request message for updating a SecurityMarks resource.

UpdateSourceRequest

Request message for updating a source.