Struct google_api_proto::google::cloud::securitycenter::v1p1beta1::security_center_client::SecurityCenterClient

pub struct SecurityCenterClient<T> { /* private fields */ }

V1p1Beta1 APIs for Security Center service.

Implementations

impl<T> SecurityCenterClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> SecurityCenterClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_source( &mut self, request: impl IntoRequest<CreateSourceRequest> ) -> Result<Response<Source>, Status>

Creates a source.

pub async fn create_finding( &mut self, request: impl IntoRequest<CreateFindingRequest> ) -> Result<Response<Finding>, Status>

Creates a finding. The corresponding source must exist for finding creation to succeed.

pub async fn create_notification_config( &mut self, request: impl IntoRequest<CreateNotificationConfigRequest> ) -> Result<Response<NotificationConfig>, Status>

Creates a notification config.

pub async fn delete_notification_config( &mut self, request: impl IntoRequest<DeleteNotificationConfigRequest> ) -> Result<Response<()>, Status>

Deletes a notification config.

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Gets the access control policy on the specified Source.

pub async fn get_notification_config( &mut self, request: impl IntoRequest<GetNotificationConfigRequest> ) -> Result<Response<NotificationConfig>, Status>

Gets a notification config.

pub async fn get_organization_settings( &mut self, request: impl IntoRequest<GetOrganizationSettingsRequest> ) -> Result<Response<OrganizationSettings>, Status>

Gets the settings for an organization.

pub async fn get_source( &mut self, request: impl IntoRequest<GetSourceRequest> ) -> Result<Response<Source>, Status>

Gets a source.

pub async fn group_assets( &mut self, request: impl IntoRequest<GroupAssetsRequest> ) -> Result<Response<GroupAssetsResponse>, Status>

Filters an organization’s assets and groups them by their specified properties.

pub async fn group_findings( &mut self, request: impl IntoRequest<GroupFindingsRequest> ) -> Result<Response<GroupFindingsResponse>, Status>

Filters an organization or source’s findings and groups them by their specified properties.

To group across all sources provide a - as the source id. Example: /v1/organizations/{organization_id}/sources/-/findings, /v1/folders/{folder_id}/sources/-/findings, /v1/projects/{project_id}/sources/-/findings

pub async fn list_assets( &mut self, request: impl IntoRequest<ListAssetsRequest> ) -> Result<Response<ListAssetsResponse>, Status>

Lists an organization’s assets.

pub async fn list_findings( &mut self, request: impl IntoRequest<ListFindingsRequest> ) -> Result<Response<ListFindingsResponse>, Status>

Lists an organization or source’s findings.

To list across all sources provide a - as the source id. Example: /v1p1beta1/organizations/{organization_id}/sources/-/findings

pub async fn list_notification_configs( &mut self, request: impl IntoRequest<ListNotificationConfigsRequest> ) -> Result<Response<ListNotificationConfigsResponse>, Status>

Lists notification configs.

pub async fn list_sources( &mut self, request: impl IntoRequest<ListSourcesRequest> ) -> Result<Response<ListSourcesResponse>, Status>

Lists all sources belonging to an organization.

pub async fn run_asset_discovery( &mut self, request: impl IntoRequest<RunAssetDiscoveryRequest> ) -> Result<Response<Operation>, Status>

Runs asset discovery. The discovery is tracked with a long-running operation.

This API can only be called with limited frequency for an organization. If it is called too frequently the caller will receive a TOO_MANY_REQUESTS error.

pub async fn set_finding_state( &mut self, request: impl IntoRequest<SetFindingStateRequest> ) -> Result<Response<Finding>, Status>

Updates the state of a finding.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Sets the access control policy on the specified Source.

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest> ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the permissions that a caller has on the specified source.

pub async fn update_finding( &mut self, request: impl IntoRequest<UpdateFindingRequest> ) -> Result<Response<Finding>, Status>

Creates or updates a finding. The corresponding source must exist for a finding creation to succeed.

pub async fn update_notification_config( &mut self, request: impl IntoRequest<UpdateNotificationConfigRequest> ) -> Result<Response<NotificationConfig>, Status>

Updates a notification config. The following update fields are allowed: description, pubsub_topic, streaming_config.filter

pub async fn update_organization_settings( &mut self, request: impl IntoRequest<UpdateOrganizationSettingsRequest> ) -> Result<Response<OrganizationSettings>, Status>

Updates an organization’s settings.

pub async fn update_source( &mut self, request: impl IntoRequest<UpdateSourceRequest> ) -> Result<Response<Source>, Status>

Updates a source.

pub async fn update_security_marks( &mut self, request: impl IntoRequest<UpdateSecurityMarksRequest> ) -> Result<Response<SecurityMarks>, Status>

Updates security marks.