Enum google_api_proto::google::cloud::securitycenter::v2::mitre_attack::Technique

#[repr(i32)]
pub enum Technique {

    Unspecified = 0,
    Masquerading = 49,
    MatchLegitimateNameOrLocation = 50,
    BootOrLogonInitializationScripts = 37,
    StartupItems = 38,
    NetworkServiceDiscovery = 32,
    ProcessDiscovery = 56,
    CommandAndScriptingInterpreter = 6,
    UnixShell = 7,
    Python = 59,
    ExploitationForPrivilegeEscalation = 63,
    PermissionGroupsDiscovery = 18,
    CloudGroups = 19,
    IndicatorRemovalFileDeletion = 64,
    ApplicationLayerProtocol = 45,
    Dns = 46,
    SoftwareDeploymentTools = 47,
    ValidAccounts = 14,
    DefaultAccounts = 35,
    LocalAccounts = 15,
    CloudAccounts = 16,
    Proxy = 9,
    ExternalProxy = 10,
    MultiHopProxy = 11,
    AccountManipulation = 22,
    AdditionalCloudCredentials = 40,
    SshAuthorizedKeys = 23,
    AdditionalContainerClusterRoles = 58,
    IngressToolTransfer = 3,
    NativeApi = 4,
    BruteForce = 44,
    SharedModules = 5,
    AccessTokenManipulation = 33,
    TokenImpersonationOrTheft = 39,
    ExploitPublicFacingApplication = 27,
    DomainPolicyModification = 30,
    DataDestruction = 29,
    ServiceStop = 52,
    InhibitSystemRecovery = 36,
    ResourceHijacking = 8,
    NetworkDenialOfService = 17,
    CloudServiceDiscovery = 48,
    StealApplicationAccessToken = 42,
    AccountAccessRemoval = 51,
    StealWebSessionCookie = 25,
    CreateOrModifySystemProcess = 24,
    AbuseElevationControlMechanism = 34,
    UnsecuredCredentials = 13,
    ModifyAuthenticationProcess = 28,
    ImpairDefenses = 31,
    DisableOrModifyTools = 55,
    ExfiltrationOverWebService = 20,
    ExfiltrationToCloudStorage = 21,
    DynamicResolution = 12,
    LateralToolTransfer = 41,
    ModifyCloudComputeInfrastructure = 26,
    CreateSnapshot = 54,
    CloudInfrastructureDiscovery = 53,
    ObtainCapabilities = 43,
    ActiveScanning = 1,
    ScanningIpBlocks = 2,
    ContainerAdministrationCommand = 60,
    EscapeToHost = 61,
    ContainerAndResourceDiscovery = 57,
    StealOrForgeAuthenticationCertificates = 62,

}

MITRE ATT&CK techniques that can be referenced by SCC findings. See: https://attack.mitre.org/techniques/enterprise/ Next ID: 65

Variants

Unspecified = 0

Unspecified value.

Masquerading = 49

T1036

MatchLegitimateNameOrLocation = 50

T1036.005

BootOrLogonInitializationScripts = 37

T1037

StartupItems = 38

T1037.005

NetworkServiceDiscovery = 32

T1046

ProcessDiscovery = 56

T1057

CommandAndScriptingInterpreter = 6

T1059

UnixShell = 7

T1059.004

Python = 59

T1059.006

ExploitationForPrivilegeEscalation = 63

T1068

PermissionGroupsDiscovery = 18

T1069

CloudGroups = 19

T1069.003

IndicatorRemovalFileDeletion = 64

T1070.004

ApplicationLayerProtocol = 45

T1071

Dns = 46

T1071.004

SoftwareDeploymentTools = 47

T1072

ValidAccounts = 14

T1078

DefaultAccounts = 35

T1078.001

LocalAccounts = 15

T1078.003

CloudAccounts = 16

T1078.004

Proxy = 9

T1090

ExternalProxy = 10

T1090.002

MultiHopProxy = 11

T1090.003

AccountManipulation = 22

T1098

AdditionalCloudCredentials = 40

T1098.001

SshAuthorizedKeys = 23

T1098.004

AdditionalContainerClusterRoles = 58

T1098.006

IngressToolTransfer = 3

T1105

NativeApi = 4

T1106

BruteForce = 44

T1110

SharedModules = 5

T1129

AccessTokenManipulation = 33

T1134

TokenImpersonationOrTheft = 39

T1134.001

ExploitPublicFacingApplication = 27

T1190

DomainPolicyModification = 30

T1484

DataDestruction = 29

T1485

ServiceStop = 52

T1489

InhibitSystemRecovery = 36

T1490

ResourceHijacking = 8

T1496

NetworkDenialOfService = 17

T1498

CloudServiceDiscovery = 48

T1526

StealApplicationAccessToken = 42

T1528

AccountAccessRemoval = 51

T1531

StealWebSessionCookie = 25

T1539

CreateOrModifySystemProcess = 24

T1543

AbuseElevationControlMechanism = 34

T1548

UnsecuredCredentials = 13

T1552

ModifyAuthenticationProcess = 28

T1556

ImpairDefenses = 31

T1562

DisableOrModifyTools = 55

T1562.001

ExfiltrationOverWebService = 20

T1567

ExfiltrationToCloudStorage = 21

T1567.002

DynamicResolution = 12

T1568

LateralToolTransfer = 41

T1570

ModifyCloudComputeInfrastructure = 26

T1578

CreateSnapshot = 54

T1578.001

CloudInfrastructureDiscovery = 53

T1580

ObtainCapabilities = 43

T1588

ActiveScanning = 1

T1595

ScanningIpBlocks = 2

T1595.001

ContainerAdministrationCommand = 60

T1609

EscapeToHost = 61

T1611

ContainerAndResourceDiscovery = 57

T1613

StealOrForgeAuthenticationCertificates = 62

T1649

Implementations

impl Technique

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Technique.

pub fn from_i32(value: i32) -> Option<Technique>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Technique, or None if value is not a valid variant.

impl Technique

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.