Struct google_api_proto::google::cloud::securitycentermanagement::v1::security_center_management_client::SecurityCenterManagementClient

pub struct SecurityCenterManagementClient<T> { /* private fields */ }

Service describing handlers for resources

Implementations

impl<T> SecurityCenterManagementClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> SecurityCenterManagementClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_effective_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListEffectiveSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListEffectiveSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all EffectiveSecurityHealthAnalyticsCustomModules for the given parent. This includes resident modules defined at the scope of the parent, and inherited modules, inherited from CRM ancestors (no descendants).

pub async fn get_effective_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<GetEffectiveSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<EffectiveSecurityHealthAnalyticsCustomModule>, Status>

Gets details of a single EffectiveSecurityHealthAnalyticsCustomModule.

pub async fn list_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all SecurityHealthAnalyticsCustomModules for the given parent. This includes resident modules defined at the scope of the parent, and inherited modules, inherited from CRM ancestors (no descendants).

pub async fn list_descendant_security_health_analytics_custom_modules( &mut self, request: impl IntoRequest<ListDescendantSecurityHealthAnalyticsCustomModulesRequest>, ) -> Result<Response<ListDescendantSecurityHealthAnalyticsCustomModulesResponse>, Status>

Returns a list of all resident SecurityHealthAnalyticsCustomModules under the given CRM parent and all of the parent’s CRM descendants.

pub async fn get_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<GetSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Retrieves a SecurityHealthAnalyticsCustomModule.

pub async fn create_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<CreateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Creates a resident SecurityHealthAnalyticsCustomModule at the scope of the given CRM parent, and also creates inherited SecurityHealthAnalyticsCustomModules for all CRM descendants of the given parent. These modules are enabled by default.

pub async fn update_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<UpdateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SecurityHealthAnalyticsCustomModule>, Status>

Updates the SecurityHealthAnalyticsCustomModule under the given name based on the given update mask. Updating the enablement state is supported on both resident and inherited modules (though resident modules cannot have an enablement state of “inherited”). Updating the display name and custom config of a module is supported on resident modules only.

pub async fn delete_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<DeleteSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<()>, Status>

Deletes the specified SecurityHealthAnalyticsCustomModule and all of its descendants in the CRM hierarchy. This method is only supported for resident custom modules.

pub async fn simulate_security_health_analytics_custom_module( &mut self, request: impl IntoRequest<SimulateSecurityHealthAnalyticsCustomModuleRequest>, ) -> Result<Response<SimulateSecurityHealthAnalyticsCustomModuleResponse>, Status>

Simulates a given SecurityHealthAnalyticsCustomModule and Resource.

pub async fn list_effective_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListEffectiveEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListEffectiveEventThreatDetectionCustomModulesResponse>, Status>

Lists all effective Event Threat Detection custom modules for the given parent. This includes resident modules defined at the scope of the parent along with modules inherited from its ancestors.

pub async fn get_effective_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<GetEffectiveEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EffectiveEventThreatDetectionCustomModule>, Status>

Gets an effective ETD custom module. Retrieves the effective module at the given level. The difference between an EffectiveCustomModule and a CustomModule is that the fields for an EffectiveCustomModule are computed from ancestors if needed. For example, the enablement_state for a CustomModule can be either ENABLED, DISABLED, or INHERITED. Where as the enablement_state for an EffectiveCustomModule is always computed to ENABLED or DISABLED (the effective enablement_state).

pub async fn list_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListEventThreatDetectionCustomModulesResponse>, Status>

Lists all Event Threat Detection custom modules for the given Resource Manager parent. This includes resident modules defined at the scope of the parent along with modules inherited from ancestors.

pub async fn list_descendant_event_threat_detection_custom_modules( &mut self, request: impl IntoRequest<ListDescendantEventThreatDetectionCustomModulesRequest>, ) -> Result<Response<ListDescendantEventThreatDetectionCustomModulesResponse>, Status>

Lists all resident Event Threat Detection custom modules under the given Resource Manager parent and its descendants.

pub async fn get_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<GetEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Gets an Event Threat Detection custom module.

pub async fn create_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<CreateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Creates a resident Event Threat Detection custom module at the scope of the given Resource Manager parent, and also creates inherited custom modules for all descendants of the given parent. These modules are enabled by default.

pub async fn update_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<UpdateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<EventThreatDetectionCustomModule>, Status>

Updates the Event Threat Detection custom module with the given name based on the given update mask. Updating the enablement state is supported for both resident and inherited modules (though resident modules cannot have an enablement state of “inherited”). Updating the display name or configuration of a module is supported for resident modules only. The type of a module cannot be changed.

pub async fn delete_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<DeleteEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<()>, Status>

Deletes the specified Event Threat Detection custom module and all of its descendants in the Resource Manager hierarchy. This method is only supported for resident custom modules.

pub async fn validate_event_threat_detection_custom_module( &mut self, request: impl IntoRequest<ValidateEventThreatDetectionCustomModuleRequest>, ) -> Result<Response<ValidateEventThreatDetectionCustomModuleResponse>, Status>

Validates the given Event Threat Detection custom module.

pub async fn get_security_center_service( &mut self, request: impl IntoRequest<GetSecurityCenterServiceRequest>, ) -> Result<Response<SecurityCenterService>, Status>

Gets service settings for the specified Security Command Center service.

pub async fn list_security_center_services( &mut self, request: impl IntoRequest<ListSecurityCenterServicesRequest>, ) -> Result<Response<ListSecurityCenterServicesResponse>, Status>

Returns a list of all Security Command Center services for the given parent.

pub async fn update_security_center_service( &mut self, request: impl IntoRequest<UpdateSecurityCenterServiceRequest>, ) -> Result<Response<SecurityCenterService>, Status>

Updates a Security Command Center service using the given update mask.