Struct google_api_proto::google::cloud::securityposture::v1::security_posture_client::SecurityPostureClient

pub struct SecurityPostureClient<T> { /* private fields */ }

Service describing handlers for resources.

Implementations

impl<T> SecurityPostureClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> SecurityPostureClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_postures( &mut self, request: impl IntoRequest<ListPosturesRequest> ) -> Result<Response<ListPosturesResponse>, Status>

(– This option restricts the visibility of the API to only projects that will (– be labeled as PREVIEW or GOOGLE_INTERNAL by the service. (– option (google.api.api_visibility).restriction = “PREVIEW,GOOGLE_INTERNAL”; Postures Lists Postures in a given organization and location. In case a posture has multiple revisions, the latest revision as per UpdateTime will be returned.

pub async fn list_posture_revisions( &mut self, request: impl IntoRequest<ListPostureRevisionsRequest> ) -> Result<Response<ListPostureRevisionsResponse>, Status>

Lists revisions of a Posture in a given organization and location.

pub async fn get_posture( &mut self, request: impl IntoRequest<GetPostureRequest> ) -> Result<Response<Posture>, Status>

Gets a posture in a given organization and location. User must provide revision_id to retrieve a specific revision of the resource. NOT_FOUND error is returned if the revision_id or the Posture name does not exist. In case revision_id is not provided then the latest Posture revision by UpdateTime is returned.

pub async fn create_posture( &mut self, request: impl IntoRequest<CreatePostureRequest> ) -> Result<Response<Operation>, Status>

Creates a new Posture resource. If a Posture with the specified name already exists in the specified organization and location, the long running operation returns a [ALREADY_EXISTS][google.rpc.Code.ALREADY_EXISTS] error.

pub async fn update_posture( &mut self, request: impl IntoRequest<UpdatePostureRequest> ) -> Result<Response<Operation>, Status>

Updates an existing Posture. A new revision of the posture will be created if the revision to be updated is currently deployed on a workload. Returns a google.rpc.Status with google.rpc.Code.NOT_FOUND if the Posture does not exist. Returns a google.rpc.Status with google.rpc.Code.ABORTED if the etag supplied in the request does not match the persisted etag of the Posture. Updatable fields are state, description and policy_sets. State update operation cannot be clubbed with update of description and policy_sets. An ACTIVE posture can be updated to both DRAFT or DEPRECATED states. Postures in DRAFT or DEPRECATED states can only be updated to ACTIVE state.

pub async fn delete_posture( &mut self, request: impl IntoRequest<DeletePostureRequest> ) -> Result<Response<Operation>, Status>

Deletes all the revisions of a resource. A posture can only be deleted when none of the revisions are deployed to any workload.

pub async fn extract_posture( &mut self, request: impl IntoRequest<ExtractPostureRequest> ) -> Result<Response<Operation>, Status>

Extracts existing policies on a workload as a posture. If a Posture on the given workload already exists, the long running operation returns a [ALREADY_EXISTS][google.rpc.Code.ALREADY_EXISTS] error.

pub async fn list_posture_deployments( &mut self, request: impl IntoRequest<ListPostureDeploymentsRequest> ) -> Result<Response<ListPostureDeploymentsResponse>, Status>

PostureDeployments Lists PostureDeployments in a given project and location.

pub async fn get_posture_deployment( &mut self, request: impl IntoRequest<GetPostureDeploymentRequest> ) -> Result<Response<PostureDeployment>, Status>

Gets details of a single PostureDeployment.

pub async fn create_posture_deployment( &mut self, request: impl IntoRequest<CreatePostureDeploymentRequest> ) -> Result<Response<Operation>, Status>

Creates a new PostureDeployment in a given project and location.

pub async fn update_posture_deployment( &mut self, request: impl IntoRequest<UpdatePostureDeploymentRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single PostureDeployment.

pub async fn delete_posture_deployment( &mut self, request: impl IntoRequest<DeletePostureDeploymentRequest> ) -> Result<Response<Operation>, Status>

Deletes a single PostureDeployment.

pub async fn list_posture_templates( &mut self, request: impl IntoRequest<ListPostureTemplatesRequest> ) -> Result<Response<ListPostureTemplatesResponse>, Status>

PostureTemplates Lists all the PostureTemplates available to the user.

pub async fn get_posture_template( &mut self, request: impl IntoRequest<GetPostureTemplateRequest> ) -> Result<Response<PostureTemplate>, Status>

Gets a PostureTemplate. User must provide revision_id to retrieve a specific revision of the resource. NOT_FOUND error is returned if the revision_id or the PostureTemplate name does not exist. In case revision_id is not provided then the PostureTemplate with latest revision_id is returned.