Struct google_api_proto::google::cloud::servicehealth::logging::v1::EventLog

pub struct EventLog {

    pub title: String,
    pub description: String,
    pub category: i32,
    pub state: i32,
    pub detailed_state: i32,
    pub impacted_products: String,
    pub impacted_locations: String,
    pub relevance: i32,
    pub parent_event: String,
    pub update_time: Option<Timestamp>,
    pub start_time: Option<Timestamp>,
    pub end_time: Option<Timestamp>,
    pub next_update_time: Option<Timestamp>,

}

Message describing the payload of service health logs.

Fields
title: String

Brief description for the event.

description: String

Free-form, human-readable description.

category: i32

Identifies the category of the event.

state: i32

The current state of the event.

detailed_state: i32

The current detailed state of the event.

impacted_products: String

Google Cloud products known to be affected by the event, in JSON serialized format. List of all supported Google Cloud products.

Example: “\['Google Cloud SQL', 'Cloud Compute Engine'\]”.

impacted_locations: String

Locations known to be impacted by the event, in JSON serialized format. See possible values, which are subject to where the service is running.

Example: “\['us-central1', 'us-west1'\]”.

relevance: i32

Communicates why a given event is deemed relevant in the context of a given project.

parent_event: String

When detailed_state=MERGED, parent_event contains the name of the parent event. All further updates will be published to the parent event.

update_time: Option<Timestamp>

The time when the event was last modified.

start_time: Option<Timestamp>

The start time of the event, if applicable.

end_time: Option<Timestamp>

The end time of the event, if applicable.

next_update_time: Option<Timestamp>

Incident-only field. The time when the next update can be expected.

Implementations

impl EventLog

pub fn category(&self) -> EventCategory

Returns the enum value of category, or the default if the field is set to an invalid enum value.

pub fn set_category(&mut self, value: EventCategory)

Sets category to the provided enum value.

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.

pub fn relevance(&self) -> Relevance

Returns the enum value of relevance, or the default if the field is set to an invalid enum value.

pub fn set_relevance(&mut self, value: Relevance)

Sets relevance to the provided enum value.

pub fn detailed_state(&self) -> DetailedState

Returns the enum value of detailed_state, or the default if the field is set to an invalid enum value.

pub fn set_detailed_state(&mut self, value: DetailedState)

Sets detailed_state to the provided enum value.