Module google_api_proto::google::cloud::servicehealth::v1

Modules

event

Nested message and enum types in Event.

organization_event

Nested message and enum types in OrganizationEvent.

service_health_client

Generated client implementations.

Structs

Asset

Represents the asset impacted by the events.

Event

Represents service health events that may affect Google Cloud products. Event resource is a read-only view and does not allow any modifications. All fields are output only.

EventImpact

Represents the Google Cloud products and locations impacted by the event.

EventUpdate

Records an update made to the event.

GetEventRequest

Message for getting an event

GetOrganizationEventRequest

GetOrganizationImpactRequest

ListEventsRequest

ListEventsResponse

ListOrganizationEventsRequest

ListOrganizationEventsResponse

ListOrganizationImpactsRequest

Message for requesting list of OrganizationImpacts

ListOrganizationImpactsResponse

Location

Represents the locations impacted by the event.

OrganizationEvent

Represents service health events that may affect Google Cloud products used across the organization. It is a read-only view and does not allow any modifications.

OrganizationImpact

Represents impact to assets at organizational level. It is a read-only view and does not allow any modifications.

Product

Represents the Google Cloud product impacted by the event.

Enums

EventView

The event fields to include in ListEvents API response. This enum lists all possible event views.

OrganizationEventView

The organization event fields to include in ListOrganizationEvents API response. This enum lists all possible organization event views.