Struct google_api_proto::google::cloud::speech::v1p1beta1::adaptation_client::AdaptationClient

pub struct AdaptationClient<T> { /* private fields */ }

Service that implements Google Cloud Speech Adaptation API.

Implementations

impl<T> AdaptationClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> AdaptationClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_phrase_set( &mut self, request: impl IntoRequest<CreatePhraseSetRequest>, ) -> Result<Response<PhraseSet>, Status>

Create a set of phrase hints. Each item in the set can be a single word or a multi-word phrase. The items in the PhraseSet are favored by the recognition model when you send a call that includes the PhraseSet.

pub async fn get_phrase_set( &mut self, request: impl IntoRequest<GetPhraseSetRequest>, ) -> Result<Response<PhraseSet>, Status>

Get a phrase set.

pub async fn list_phrase_set( &mut self, request: impl IntoRequest<ListPhraseSetRequest>, ) -> Result<Response<ListPhraseSetResponse>, Status>

List phrase sets.

pub async fn update_phrase_set( &mut self, request: impl IntoRequest<UpdatePhraseSetRequest>, ) -> Result<Response<PhraseSet>, Status>

Update a phrase set.

pub async fn delete_phrase_set( &mut self, request: impl IntoRequest<DeletePhraseSetRequest>, ) -> Result<Response<()>, Status>

Delete a phrase set.

pub async fn create_custom_class( &mut self, request: impl IntoRequest<CreateCustomClassRequest>, ) -> Result<Response<CustomClass>, Status>

Create a custom class.

pub async fn get_custom_class( &mut self, request: impl IntoRequest<GetCustomClassRequest>, ) -> Result<Response<CustomClass>, Status>

Get a custom class.

pub async fn list_custom_classes( &mut self, request: impl IntoRequest<ListCustomClassesRequest>, ) -> Result<Response<ListCustomClassesResponse>, Status>

List custom classes.

pub async fn update_custom_class( &mut self, request: impl IntoRequest<UpdateCustomClassRequest>, ) -> Result<Response<CustomClass>, Status>

Update a custom class.

pub async fn delete_custom_class( &mut self, request: impl IntoRequest<DeleteCustomClassRequest>, ) -> Result<Response<()>, Status>

Delete a custom class.