Enum google_api_proto::google::cloud::sql::v1::operation::SqlOperationType

#[repr(i32)]
pub enum SqlOperationType {

    Unspecified = 0,
    Import = 1,
    Export = 2,
    Create = 3,
    Update = 4,
    Delete = 5,
    Restart = 6,
    Backup = 7,
    Snapshot = 8,
    BackupVolume = 9,
    DeleteVolume = 10,
    RestoreVolume = 11,
    InjectUser = 12,
    Clone = 14,
    StopReplica = 15,
    StartReplica = 16,
    PromoteReplica = 17,
    CreateReplica = 18,
    CreateUser = 19,
    DeleteUser = 20,
    UpdateUser = 21,
    CreateDatabase = 22,
    DeleteDatabase = 23,
    UpdateDatabase = 24,
    Failover = 25,
    DeleteBackup = 26,
    RecreateReplica = 27,
    TruncateLog = 28,
    DemoteMaster = 29,
    Maintenance = 30,
    EnablePrivateIp = 31,
    DeferMaintenance = 32,
    CreateClone = 33,
    RescheduleMaintenance = 34,
    StartExternalSync = 35,
    LogCleanup = 36,
    AutoRestart = 37,
    Reencrypt = 38,
    Switchover = 39,

}

The type of Cloud SQL operation.

Variants

Unspecified = 0

Unknown operation type.

Import = 1

Imports data into a Cloud SQL instance.

Export = 2

Exports data from a Cloud SQL instance to a Cloud Storage bucket.

Create = 3

Creates a new Cloud SQL instance.

Update = 4

Updates the settings of a Cloud SQL instance.

Delete = 5

Deletes a Cloud SQL instance.

Restart = 6

Restarts the Cloud SQL instance.

Backup = 7

Snapshot = 8

BackupVolume = 9

Performs instance backup.

DeleteVolume = 10

Deletes an instance backup.

RestoreVolume = 11

Restores an instance backup.

InjectUser = 12

Injects a privileged user in mysql for MOB instances.

Clone = 14

Clones a Cloud SQL instance.

StopReplica = 15

Stops replication on a Cloud SQL read replica instance.

StartReplica = 16

Starts replication on a Cloud SQL read replica instance.

PromoteReplica = 17

Promotes a Cloud SQL replica instance.

CreateReplica = 18

Creates a Cloud SQL replica instance.

CreateUser = 19

Creates a new user in a Cloud SQL instance.

DeleteUser = 20

Deletes a user from a Cloud SQL instance.

UpdateUser = 21

Updates an existing user in a Cloud SQL instance.

CreateDatabase = 22

Creates a database in the Cloud SQL instance.

DeleteDatabase = 23

Deletes a database in the Cloud SQL instance.

UpdateDatabase = 24

Updates a database in the Cloud SQL instance.

Failover = 25

Performs failover of an HA-enabled Cloud SQL failover replica.

DeleteBackup = 26

Deletes the backup taken by a backup run.

RecreateReplica = 27

TruncateLog = 28

Truncates a general or slow log table in MySQL.

DemoteMaster = 29

Demotes the stand-alone instance to be a Cloud SQL read replica for an external database server.

Maintenance = 30

Indicates that the instance is currently in maintenance. Maintenance typically causes the instance to be unavailable for 1-3 minutes.

EnablePrivateIp = 31

This field is deprecated, and will be removed in future version of API.

DeferMaintenance = 32

CreateClone = 33

Creates clone instance.

RescheduleMaintenance = 34

Reschedule maintenance to another time.

StartExternalSync = 35

Starts external sync of a Cloud SQL EM replica to an external primary instance.

LogCleanup = 36

Recovers logs from an instanceโs old data disk.

AutoRestart = 37

Performs auto-restart of an HA-enabled Cloud SQL database for auto recovery.

Reencrypt = 38

Re-encrypts CMEK instances with latest key version.

Switchover = 39

Switches over to replica instance from primary.

Implementations

impl SqlOperationType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of SqlOperationType.

pub fn from_i32(value: i32) -> Option<SqlOperationType>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a SqlOperationType, or None if value is not a valid variant.

impl SqlOperationType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.