Struct google_api_proto::google::cloud::sql::v1beta4::DatabaseInstance

pub struct DatabaseInstance {

    pub kind: String,
    pub state: i32,
    pub database_version: i32,
    pub settings: Option<Settings>,
    pub etag: String,
    pub failover_replica: Option<SqlFailoverReplica>,
    pub master_instance_name: String,
    pub replica_names: Vec<String>,
    pub max_disk_size: Option<i64>,
    pub current_disk_size: Option<i64>,
    pub ip_addresses: Vec<IpMapping>,
    pub server_ca_cert: Option<SslCert>,
    pub instance_type: i32,
    pub project: String,
    pub ipv6_address: String,
    pub service_account_email_address: String,
    pub on_premises_configuration: Option<OnPremisesConfiguration>,
    pub replica_configuration: Option<ReplicaConfiguration>,
    pub backend_type: i32,
    pub self_link: String,
    pub suspension_reason: Vec<i32>,
    pub connection_name: String,
    pub name: String,
    pub region: String,
    pub gce_zone: String,
    pub secondary_gce_zone: String,
    pub disk_encryption_configuration: Option<DiskEncryptionConfiguration>,
    pub disk_encryption_status: Option<DiskEncryptionStatus>,
    pub root_password: String,
    pub scheduled_maintenance: Option<SqlScheduledMaintenance>,
    pub satisfies_pzs: Option<bool>,
    pub database_installed_version: String,
    pub out_of_disk_report: Option<SqlOutOfDiskReport>,
    pub create_time: Option<Timestamp>,
    pub available_maintenance_versions: Vec<String>,
    pub maintenance_version: String,
    pub sql_network_architecture: Option<i32>,
    pub psc_service_attachment_link: Option<String>,
    pub dns_name: Option<String>,
    pub primary_dns_name: Option<String>,
    pub write_endpoint: Option<String>,

}

A Cloud SQL instance resource.

Fields
kind: String

This is always sql#instance.

state: i32

The current serving state of the Cloud SQL instance.

database_version: i32

The database engine type and version. The databaseVersion field cannot be changed after instance creation.

settings: Option<Settings>

The user settings.

etag: String

This field is deprecated and will be removed from a future version of the API. Use the settings.settingsVersion field instead.

failover_replica: Option<SqlFailoverReplica>

The name and status of the failover replica.

master_instance_name: String

The name of the instance which will act as primary in the replication setup.

replica_names: Vec<String>

The replicas of the instance.

max_disk_size: Option<i64>
๐Deprecated

The maximum disk size of the instance in bytes.

current_disk_size: Option<i64>
๐Deprecated

The current disk usage of the instance in bytes. This property has been deprecated. Use the โcloudsql.googleapis.com/database/disk/bytes_usedโ metric in Cloud Monitoring API instead. Please see this announcement for details.

ip_addresses: Vec<IpMapping>

The assigned IP addresses for the instance.

server_ca_cert: Option<SslCert>

SSL configuration.

instance_type: i32

The instance type.

project: String

The project ID of the project containing the Cloud SQL instance. The Google apps domain is prefixed if applicable.

ipv6_address: String
๐Deprecated

The IPv6 address assigned to the instance. (Deprecated) This property was applicable only to First Generation instances.

service_account_email_address: String

The service account email address assigned to the instance. \This property is read-only.

on_premises_configuration: Option<OnPremisesConfiguration>

Configuration specific to on-premises instances.

replica_configuration: Option<ReplicaConfiguration>

Configuration specific to failover replicas and read replicas.

backend_type: i32

The backend type. SECOND_GEN: Cloud SQL database instance. EXTERNAL: A database server that is not managed by Google.

This property is read-only; use the tier property in the settings object to determine the database type.

self_link: String

The URI of this resource.

suspension_reason: Vec<i32>

If the instance state is SUSPENDED, the reason for the suspension.

connection_name: String

Connection name of the Cloud SQL instance used in connection strings.

name: String

Name of the Cloud SQL instance. This does not include the project ID.

region: String

The geographical region. Can be:

us-central (FIRST_GEN instances only)

us-central1 (SECOND_GEN instances only)

asia-east1 or europe-west1.

Defaults to us-central or us-central1 depending on the instance type. The region cannot be changed after instance creation.

gce_zone: String

The Compute Engine zone that the instance is currently serving from. This value could be different from the zone that was specified when the instance was created if the instance has failed over to its secondary zone. WARNING: Changing this might restart the instance.

secondary_gce_zone: String

The Compute Engine zone that the failover instance is currently serving from for a regional instance. This value could be different from the zone that was specified when the instance was created if the instance has failed over to its secondary/failover zone.

disk_encryption_configuration: Option<DiskEncryptionConfiguration>

Disk encryption configuration specific to an instance.

disk_encryption_status: Option<DiskEncryptionStatus>

Disk encryption status specific to an instance.

root_password: String

Initial root password. Use only on creation. You must set root passwords before you can connect to PostgreSQL instances.

scheduled_maintenance: Option<SqlScheduledMaintenance>

The start time of any upcoming scheduled maintenance for this instance.

satisfies_pzs: Option<bool>

The status indicating if instance satisfiesPzs. Reserved for future use.

database_installed_version: String

Output only. Stores the current database version running on the instance including minor version such as MYSQL_8_0_18.

out_of_disk_report: Option<SqlOutOfDiskReport>

This field represents the report generated by the proactive database wellness job for OutOfDisk issues.

Writers:

the proactive database wellness job for OOD.

Readers:

the proactive database wellness job

create_time: Option<Timestamp>

Output only. The time when the instance was created in RFC 3339 format, for example 2012-11-15T16:19:00.094Z.

available_maintenance_versions: Vec<String>

Output only. List all maintenance versions applicable on the instance

maintenance_version: String

The current software version on the instance.

sql_network_architecture: Option<i32>

The SQL network architecture for the instance.

psc_service_attachment_link: Option<String>

Output only. The link to service attachment of PSC instance.

dns_name: Option<String>

Output only. The dns name of the instance.

primary_dns_name: Option<String>
๐Deprecated

Output only. DEPRECATED: please use write_endpoint instead.

write_endpoint: Option<String>

Output only. The dns name of the primary instance in a replication group.

Implementations

impl DatabaseInstance

pub fn state(&self) -> SqlInstanceState

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: SqlInstanceState)

Sets state to the provided enum value.

pub fn database_version(&self) -> SqlDatabaseVersion

Returns the enum value of database_version, or the default if the field is set to an invalid enum value.

pub fn set_database_version(&mut self, value: SqlDatabaseVersion)

Sets database_version to the provided enum value.

pub fn instance_type(&self) -> SqlInstanceType

Returns the enum value of instance_type, or the default if the field is set to an invalid enum value.

pub fn set_instance_type(&mut self, value: SqlInstanceType)

Sets instance_type to the provided enum value.

pub fn backend_type(&self) -> SqlBackendType

Returns the enum value of backend_type, or the default if the field is set to an invalid enum value.

pub fn set_backend_type(&mut self, value: SqlBackendType)

Sets backend_type to the provided enum value.

pub fn suspension_reason( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<SqlSuspensionReason>>

Returns an iterator which yields the valid enum values contained in suspension_reason.

pub fn push_suspension_reason(&mut self, value: SqlSuspensionReason)

Appends the provided enum value to suspension_reason.

pub fn sql_network_architecture(&self) -> SqlNetworkArchitecture

Returns the enum value of sql_network_architecture, or the default if the field is unset or set to an invalid enum value.

pub fn set_sql_network_architecture(&mut self, value: SqlNetworkArchitecture)

Sets sql_network_architecture to the provided enum value.

pub fn psc_service_attachment_link(&self) -> &str

Returns the value of psc_service_attachment_link, or the default value if psc_service_attachment_link is unset.

pub fn dns_name(&self) -> &str

Returns the value of dns_name, or the default value if dns_name is unset.

pub fn primary_dns_name(&self) -> &str

Returns the value of primary_dns_name, or the default value if primary_dns_name is unset.

pub fn write_endpoint(&self) -> &str

Returns the value of write_endpoint, or the default value if write_endpoint is unset.