Module google_api_proto::google::cloud::storageinsights::v1

Modules

frequency_options

Nested message and enum types in FrequencyOptions.

object_metadata_report_options

Nested message and enum types in ObjectMetadataReportOptions.

report_config

Nested message and enum types in ReportConfig.

report_detail

Nested message and enum types in ReportDetail.

storage_insights_client

Generated client implementations.

Structs

CloudStorageDestinationOptions

Options to store reports in storage systems. Next ID: 3

CloudStorageFilters

Options to filter data on storage systems. Next ID: 2

CreateReportConfigRequest

Message for creating a ReportConfig

CsvOptions

Options to configure CSV formatted reports.

DeleteReportConfigRequest

Message for deleting a ReportConfig

FrequencyOptions

ReportConfig Resource:

GetReportConfigRequest

Message for getting a ReportConfig

GetReportDetailRequest

Message for getting a ReportDetail

ListReportConfigsRequest

Message for requesting list of ReportConfigs

ListReportConfigsResponse

Message for response to listing ReportConfigs

ListReportDetailsRequest

Message for requesting list of ReportDetails

ListReportDetailsResponse

Message for response to listing ReportDetails

ObjectMetadataReportOptions

Report specification for exporting object metadata. Next ID: 4

OperationMetadata

Represents the metadata of the long-running operation.

ParquetOptions

Options to configure Parquet formatted reports.

ReportConfig

Message describing ReportConfig object. ReportConfig is the configuration to generate reports. See https://cloud.google.com/storage/docs/insights/using-inventory-reports#create-config-rest for more details on how to set various fields. Next ID: 12

ReportDetail

Message describing ReportDetail object. ReportDetail represents metadata of generated reports for a ReportConfig. Next ID: 10

UpdateReportConfigRequest

Message for updating a ReportConfig