Enum google_api_proto::google::cloud::talent::v4::JobCategory

#[repr(i32)]
pub enum JobCategory {

    Unspecified,
    AccountingAndFinance,
    AdministrativeAndOffice,
    AdvertisingAndMarketing,
    AnimalCare,
    ArtFashionAndDesign,
    BusinessOperations,
    CleaningAndFacilities,
    ComputerAndIt,
    Construction,
    CustomerService,
    Education,
    EntertainmentAndTravel,
    FarmingAndOutdoors,
    Healthcare,
    HumanResources,
    InstallationMaintenanceAndRepair,
    Legal,
    Management,
    ManufacturingAndWarehouse,
    MediaCommunicationsAndWriting,
    OilGasAndMining,
    PersonalCareAndServices,
    ProtectiveServices,
    RealEstate,
    RestaurantAndHospitality,
    SalesAndRetail,
    ScienceAndEngineering,
    SocialServicesAndNonProfit,
    SportsFitnessAndRecreation,
    TransportationAndLogistics,

}

An enum that represents the categorization or primary focus of specific role. This value is different than the “industry” associated with a role, which is related to the categorization of the company listing the job.

Variants

Unspecified

The default value if the category isn’t specified.

AccountingAndFinance

An accounting and finance job, such as an Accountant.

AdministrativeAndOffice

An administrative and office job, such as an Administrative Assistant.

AdvertisingAndMarketing

An advertising and marketing job, such as Marketing Manager.

AnimalCare

An animal care job, such as Veterinarian.

ArtFashionAndDesign

An art, fashion, or design job, such as Designer.

BusinessOperations

A business operations job, such as Business Operations Manager.

CleaningAndFacilities

A cleaning and facilities job, such as Custodial Staff.

ComputerAndIt

A computer and IT job, such as Systems Administrator.

Construction

A construction job, such as General Laborer.

CustomerService

A customer service job, such s Cashier.

Education

An education job, such as School Teacher.

EntertainmentAndTravel

An entertainment and travel job, such as Flight Attendant.

FarmingAndOutdoors

A farming or outdoor job, such as Park Ranger.

Healthcare

A healthcare job, such as Registered Nurse.

HumanResources

A human resources job, such as Human Resources Director.

InstallationMaintenanceAndRepair

An installation, maintenance, or repair job, such as Electrician.

Legal

A legal job, such as Law Clerk.

Management

A management job, often used in conjunction with another category, such as Store Manager.

ManufacturingAndWarehouse

A manufacturing or warehouse job, such as Assembly Technician.

MediaCommunicationsAndWriting

A media, communications, or writing job, such as Media Relations.

OilGasAndMining

An oil, gas or mining job, such as Offshore Driller.

PersonalCareAndServices

A personal care and services job, such as Hair Stylist.

ProtectiveServices

A protective services job, such as Security Guard.

RealEstate

A real estate job, such as Buyer’s Agent.

RestaurantAndHospitality

A restaurant and hospitality job, such as Restaurant Server.

SalesAndRetail

A sales and/or retail job, such Sales Associate.

ScienceAndEngineering

A science and engineering job, such as Lab Technician.

SocialServicesAndNonProfit

A social services or non-profit job, such as Case Worker.

SportsFitnessAndRecreation

A sports, fitness, or recreation job, such as Personal Trainer.

TransportationAndLogistics

A transportation or logistics job, such as Truck Driver.

Implementations

impl JobCategory

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of JobCategory.

pub fn from_i32(value: i32) -> Option<JobCategory>

Converts an i32 to a JobCategory, or None if value is not a valid variant.