Struct google_api_proto::google::cloud::talent::v4beta1::job_service_client::JobServiceClient

pub struct JobServiceClient<T> { /* private fields */ }

A service handles job management, including job CRUD, enumeration and search.

Implementations

impl<T> JobServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> JobServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_job( &mut self, request: impl IntoRequest<CreateJobRequest> ) -> Result<Response<Job>, Status>

Creates a new job.

Typically, the job becomes searchable within 10 seconds, but it may take up to 5 minutes.

pub async fn batch_create_jobs( &mut self, request: impl IntoRequest<BatchCreateJobsRequest> ) -> Result<Response<Operation>, Status>

Begins executing a batch create jobs operation.

pub async fn get_job( &mut self, request: impl IntoRequest<GetJobRequest> ) -> Result<Response<Job>, Status>

Retrieves the specified job, whose status is OPEN or recently EXPIRED within the last 90 days.

pub async fn update_job( &mut self, request: impl IntoRequest<UpdateJobRequest> ) -> Result<Response<Job>, Status>

Updates specified job.

Typically, updated contents become visible in search results within 10 seconds, but it may take up to 5 minutes.

pub async fn batch_update_jobs( &mut self, request: impl IntoRequest<BatchUpdateJobsRequest> ) -> Result<Response<Operation>, Status>

Begins executing a batch update jobs operation.

pub async fn delete_job( &mut self, request: impl IntoRequest<DeleteJobRequest> ) -> Result<Response<()>, Status>

Deletes the specified job.

Typically, the job becomes unsearchable within 10 seconds, but it may take up to 5 minutes.

pub async fn batch_delete_jobs( &mut self, request: impl IntoRequest<BatchDeleteJobsRequest> ) -> Result<Response<()>, Status>

Deletes a list of [Job][google.cloud.talent.v4beta1.Job]s by filter.

pub async fn list_jobs( &mut self, request: impl IntoRequest<ListJobsRequest> ) -> Result<Response<ListJobsResponse>, Status>

Lists jobs by filter.

pub async fn search_jobs( &mut self, request: impl IntoRequest<SearchJobsRequest> ) -> Result<Response<SearchJobsResponse>, Status>

Searches for jobs using the provided [SearchJobsRequest][google.cloud.talent.v4beta1.SearchJobsRequest].

This call constrains the [visibility][google.cloud.talent.v4beta1.Job.visibility] of jobs present in the database, and only returns jobs that the caller has permission to search against.

pub async fn search_jobs_for_alert( &mut self, request: impl IntoRequest<SearchJobsRequest> ) -> Result<Response<SearchJobsResponse>, Status>

Searches for jobs using the provided [SearchJobsRequest][google.cloud.talent.v4beta1.SearchJobsRequest].

This API call is intended for the use case of targeting passive job seekers (for example, job seekers who have signed up to receive email alerts about potential job opportunities), and has different algorithmic adjustments that are targeted to passive job seekers.

This call constrains the [visibility][google.cloud.talent.v4beta1.Job.visibility] of jobs present in the database, and only returns jobs the caller has permission to search against.