Module google_api_proto::google::cloud::tasks::v2beta3

Modules

cloud_tasks_client

Generated client implementations.

http_request

Nested message and enum types in HttpRequest.

http_target

Nested message and enum types in HttpTarget.

queue

Nested message and enum types in Queue.

task

Nested message and enum types in Task.

uri_override

Nested message and enum types in UriOverride.

Structs

AppEngineHttpQueue

App Engine HTTP queue.

AppEngineHttpRequest

App Engine HTTP request.

AppEngineRouting

App Engine Routing.

Attempt

The status of a task attempt.

CreateQueueRequest

Request message for [CreateQueue][google.cloud.tasks.v2beta3.CloudTasks.CreateQueue].

CreateTaskRequest

Request message for [CreateTask][google.cloud.tasks.v2beta3.CloudTasks.CreateTask].

DeleteQueueRequest

Request message for [DeleteQueue][google.cloud.tasks.v2beta3.CloudTasks.DeleteQueue].

DeleteTaskRequest

Request message for deleting a task using [DeleteTask][google.cloud.tasks.v2beta3.CloudTasks.DeleteTask].

GetQueueRequest

Request message for [GetQueue][google.cloud.tasks.v2beta3.CloudTasks.GetQueue].

GetTaskRequest

Request message for getting a task using [GetTask][google.cloud.tasks.v2beta3.CloudTasks.GetTask].

HttpRequest

HTTP request.

HttpTarget

HTTP target.

ListQueuesRequest

Request message for [ListQueues][google.cloud.tasks.v2beta3.CloudTasks.ListQueues].

ListQueuesResponse

Response message for [ListQueues][google.cloud.tasks.v2beta3.CloudTasks.ListQueues].

ListTasksRequest

Request message for listing tasks using [ListTasks][google.cloud.tasks.v2beta3.CloudTasks.ListTasks].

ListTasksResponse

Response message for listing tasks using [ListTasks][google.cloud.tasks.v2beta3.CloudTasks.ListTasks].

OAuthToken

Contains information needed for generating an OAuth token. This type of authorization should generally only be used when calling Google APIs hosted on *.googleapis.com.

OidcToken

Contains information needed for generating an OpenID Connect token. This type of authorization can be used for many scenarios, including calling Cloud Run, or endpoints where you intend to validate the token yourself.

PathOverride

PathOverride.

PauseQueueRequest

Request message for [PauseQueue][google.cloud.tasks.v2beta3.CloudTasks.PauseQueue].

PullMessage

Pull Message.

PurgeQueueRequest

Request message for [PurgeQueue][google.cloud.tasks.v2beta3.CloudTasks.PurgeQueue].

QueryOverride

QueryOverride.

Queue

A queue is a container of related tasks. Queues are configured to manage how those tasks are dispatched. Configurable properties include rate limits, retry options, queue types, and others.

QueueStats

Statistics for a queue.

RateLimits

Rate limits.

ResumeQueueRequest

Request message for [ResumeQueue][google.cloud.tasks.v2beta3.CloudTasks.ResumeQueue].

RetryConfig

Retry config.

RunTaskRequest

Request message for forcing a task to run now using [RunTask][google.cloud.tasks.v2beta3.CloudTasks.RunTask].

StackdriverLoggingConfig

Configuration options for writing logs to Stackdriver Logging.

Task

A unit of scheduled work.

UpdateQueueRequest

Request message for [UpdateQueue][google.cloud.tasks.v2beta3.CloudTasks.UpdateQueue].

UriOverride

URI Override.

Enums

HttpMethod

The HTTP method used to execute the task.