Struct google_api_proto::google::cloud::tasks::v2beta3::HttpRequest

pub struct HttpRequest {
    pub url: String,
    pub http_method: i32,
    pub headers: BTreeMap<String, String>,
    pub body: Bytes,
    pub authorization_header: Option<AuthorizationHeader>,
}

HTTP request.

The task will be pushed to the worker as an HTTP request. If the worker or the redirected worker acknowledges the task by returning a successful HTTP response code ([200 - 299]), the task will be removed from the queue. If any other HTTP response code is returned or no response is received, the task will be retried according to the following:

User-specified throttling: [retry configuration][google.cloud.tasks.v2beta3.Queue.retry_config], [rate limits][google.cloud.tasks.v2beta3.Queue.rate_limits], and the [queue’s state][google.cloud.tasks.v2beta3.Queue.state].

System throttling: To prevent the worker from overloading, Cloud Tasks may temporarily reduce the queue’s effective rate. User-specified settings will not be changed.

System throttling happens because:

Cloud Tasks backs off on all errors. Normally the backoff specified in [rate limits][google.cloud.tasks.v2beta3.Queue.rate_limits] will be used. But if the worker returns 429 (Too Many Requests), 503 (Service Unavailable), or the rate of errors is high, Cloud Tasks will use a higher backoff rate. The retry specified in the Retry-After HTTP response header is considered.

To prevent traffic spikes and to smooth sudden increases in traffic, dispatches ramp up slowly when the queue is newly created or idle and if large numbers of tasks suddenly become available to dispatch (due to spikes in create task rates, the queue being unpaused, or many tasks that are scheduled at the same time).

Fields
url: String

Required. The full url path that the request will be sent to.

This string must begin with either “http://" or “https://". Some examples are: <http://acme.com> and <https://acme.com/sales:8080.> Cloud Tasks will encode some characters for safety and compatibility. The maximum allowed URL length is 2083 characters after encoding.

The Location header response from a redirect response [300 - 399] may be followed. The redirect is not counted as a separate attempt.

http_method: i32

The HTTP method to use for the request. The default is POST.

headers: BTreeMap<String, String>

HTTP request headers.

This map contains the header field names and values. Headers can be set when the [task is created][google.cloud.tasks.v2beta3.CloudTasks.CreateTask].

These headers represent a subset of the headers that will accompany the task’s HTTP request. Some HTTP request headers will be ignored or replaced.

A partial list of headers that will be ignored or replaced is:

Any header that is prefixed with “X-CloudTasks-” will be treated as service header. Service headers define properties of the task and are predefined in CloudTask.

Host: This will be computed by Cloud Tasks and derived from [HttpRequest.url][google.cloud.tasks.v2beta3.HttpRequest.url].

Content-Length: This will be computed by Cloud Tasks.

User-Agent: This will be set to "Google-Cloud-Tasks".

X-Google-*: Google use only.

X-AppEngine-*: Google use only.

Content-Type won’t be set by Cloud Tasks. You can explicitly set Content-Type to a media type when the [task is created][google.cloud.tasks.v2beta3.CloudTasks.CreateTask]. For example, Content-Type can be set to "application/octet-stream" or "application/json".

Headers which can have multiple values (according to RFC2616) can be specified using comma-separated values.

The size of the headers must be less than 80KB.

body: Bytes

HTTP request body.

A request body is allowed only if the [HTTP method][google.cloud.tasks.v2beta3.HttpRequest.http_method] is POST, PUT, or PATCH. It is an error to set body on a task with an incompatible [HttpMethod][google.cloud.tasks.v2beta3.HttpMethod].

authorization_header: Option<AuthorizationHeader>

The mode for generating an Authorization header for HTTP requests.

If specified, all Authorization headers in the [HttpRequest.headers][google.cloud.tasks.v2beta3.HttpRequest.headers] field will be overridden.

Implementations

impl HttpRequest

pub fn http_method(&self) -> HttpMethod

Returns the enum value of http_method, or the default if the field is set to an invalid enum value.

pub fn set_http_method(&mut self, value: HttpMethod)

Sets http_method to the provided enum value.