Struct google_api_proto::google::cloud::telcoautomation::v1alpha1::telco_automation_client::TelcoAutomationClient

pub struct TelcoAutomationClient<T> { /* private fields */ }

TelcoAutomation Service manages the control plane cluster a.k.a. Orchestration Cluster (GKE cluster with config controller) of TNA. It also exposes blueprint APIs which manages the lifecycle of blueprints that control the infrastructure setup (e.g GDCE clusters) and deployment of network functions.

Implementations

impl<T> TelcoAutomationClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> TelcoAutomationClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_orchestration_clusters( &mut self, request: impl IntoRequest<ListOrchestrationClustersRequest> ) -> Result<Response<ListOrchestrationClustersResponse>, Status>

Lists OrchestrationClusters in a given project and location.

pub async fn get_orchestration_cluster( &mut self, request: impl IntoRequest<GetOrchestrationClusterRequest> ) -> Result<Response<OrchestrationCluster>, Status>

Gets details of a single OrchestrationCluster.

pub async fn create_orchestration_cluster( &mut self, request: impl IntoRequest<CreateOrchestrationClusterRequest> ) -> Result<Response<Operation>, Status>

Creates a new OrchestrationCluster in a given project and location.

pub async fn delete_orchestration_cluster( &mut self, request: impl IntoRequest<DeleteOrchestrationClusterRequest> ) -> Result<Response<Operation>, Status>

Deletes a single OrchestrationCluster.

pub async fn list_edge_slms( &mut self, request: impl IntoRequest<ListEdgeSlmsRequest> ) -> Result<Response<ListEdgeSlmsResponse>, Status>

Lists EdgeSlms in a given project and location.

pub async fn get_edge_slm( &mut self, request: impl IntoRequest<GetEdgeSlmRequest> ) -> Result<Response<EdgeSlm>, Status>

Gets details of a single EdgeSlm.

pub async fn create_edge_slm( &mut self, request: impl IntoRequest<CreateEdgeSlmRequest> ) -> Result<Response<Operation>, Status>

Creates a new EdgeSlm in a given project and location.

pub async fn delete_edge_slm( &mut self, request: impl IntoRequest<DeleteEdgeSlmRequest> ) -> Result<Response<Operation>, Status>

Deletes a single EdgeSlm.

pub async fn create_blueprint( &mut self, request: impl IntoRequest<CreateBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Creates a blueprint.

pub async fn update_blueprint( &mut self, request: impl IntoRequest<UpdateBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Updates a blueprint.

pub async fn get_blueprint( &mut self, request: impl IntoRequest<GetBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Returns the requested blueprint.

pub async fn delete_blueprint( &mut self, request: impl IntoRequest<DeleteBlueprintRequest> ) -> Result<Response<()>, Status>

Deletes a blueprint and all its revisions.

pub async fn delete_blueprint_revision( &mut self, request: impl IntoRequest<DeleteBlueprintRevisionRequest> ) -> Result<Response<Blueprint>, Status>

Deletes the specified revision of the blueprint.

pub async fn list_blueprints( &mut self, request: impl IntoRequest<ListBlueprintsRequest> ) -> Result<Response<ListBlueprintsResponse>, Status>

List all blueprints.

pub async fn approve_blueprint( &mut self, request: impl IntoRequest<ApproveBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Approves a blueprint and commits a new revision.

pub async fn propose_blueprint( &mut self, request: impl IntoRequest<ProposeBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Proposes a blueprint for approval of changes.

pub async fn reject_blueprint( &mut self, request: impl IntoRequest<RejectBlueprintRequest> ) -> Result<Response<Blueprint>, Status>

Rejects a blueprint revision proposal and flips it back to Draft state.

pub async fn list_blueprint_revisions( &mut self, request: impl IntoRequest<ListBlueprintRevisionsRequest> ) -> Result<Response<ListBlueprintRevisionsResponse>, Status>

List blueprint revisions of a given blueprint.

pub async fn search_blueprint_revisions( &mut self, request: impl IntoRequest<SearchBlueprintRevisionsRequest> ) -> Result<Response<SearchBlueprintRevisionsResponse>, Status>

Searches across blueprint revisions.

pub async fn search_deployment_revisions( &mut self, request: impl IntoRequest<SearchDeploymentRevisionsRequest> ) -> Result<Response<SearchDeploymentRevisionsResponse>, Status>

Searches across deployment revisions.

pub async fn discard_blueprint_changes( &mut self, request: impl IntoRequest<DiscardBlueprintChangesRequest> ) -> Result<Response<DiscardBlueprintChangesResponse>, Status>

Discards the changes in a blueprint and reverts the blueprint to the last approved blueprint revision. No changes take place if a blueprint does not have revisions.

pub async fn list_public_blueprints( &mut self, request: impl IntoRequest<ListPublicBlueprintsRequest> ) -> Result<Response<ListPublicBlueprintsResponse>, Status>

Lists the blueprints in TNA’s public catalog. Default page size = 20, Max Page Size = 100.

pub async fn get_public_blueprint( &mut self, request: impl IntoRequest<GetPublicBlueprintRequest> ) -> Result<Response<PublicBlueprint>, Status>

Returns the requested public blueprint.

pub async fn create_deployment( &mut self, request: impl IntoRequest<CreateDeploymentRequest> ) -> Result<Response<Deployment>, Status>

Creates a deployment.

pub async fn update_deployment( &mut self, request: impl IntoRequest<UpdateDeploymentRequest> ) -> Result<Response<Deployment>, Status>

Updates a deployment.

pub async fn get_deployment( &mut self, request: impl IntoRequest<GetDeploymentRequest> ) -> Result<Response<Deployment>, Status>

Returns the requested deployment.

pub async fn delete_deployment( &mut self, request: impl IntoRequest<DeleteDeploymentRequest> ) -> Result<Response<()>, Status>

Deletes a deployment and all its revisions.

pub async fn remove_deployment( &mut self, request: impl IntoRequest<RemoveDeploymentRequest> ) -> Result<Response<()>, Status>

Removes the deployment by marking it as DELETING. Post which deployment and it’s revisions gets deleted.

pub async fn delete_deployment_revision( &mut self, request: impl IntoRequest<DeleteDeploymentRevisionRequest> ) -> Result<Response<Deployment>, Status>

Deletes the specified revision of the deployment.

pub async fn list_deployments( &mut self, request: impl IntoRequest<ListDeploymentsRequest> ) -> Result<Response<ListDeploymentsResponse>, Status>

List all deployments.

pub async fn list_deployment_revisions( &mut self, request: impl IntoRequest<ListDeploymentRevisionsRequest> ) -> Result<Response<ListDeploymentRevisionsResponse>, Status>

List deployment revisions of a given deployment.

pub async fn discard_deployment_changes( &mut self, request: impl IntoRequest<DiscardDeploymentChangesRequest> ) -> Result<Response<DiscardDeploymentChangesResponse>, Status>

Discards the changes in a deployment and reverts the deployment to the last approved deployment revision. No changes take place if a deployment does not have revisions.

pub async fn apply_deployment( &mut self, request: impl IntoRequest<ApplyDeploymentRequest> ) -> Result<Response<Deployment>, Status>

Applies the deployment’s YAML files to the parent orchestration cluster.

pub async fn compute_deployment_status( &mut self, request: impl IntoRequest<ComputeDeploymentStatusRequest> ) -> Result<Response<ComputeDeploymentStatusResponse>, Status>

Returns the requested deployment status.

pub async fn rollback_deployment( &mut self, request: impl IntoRequest<RollbackDeploymentRequest> ) -> Result<Response<Deployment>, Status>

Rollback the active deployment to the given past approved deployment revision.

pub async fn get_hydrated_deployment( &mut self, request: impl IntoRequest<GetHydratedDeploymentRequest> ) -> Result<Response<HydratedDeployment>, Status>

Returns the requested hydrated deployment.

pub async fn list_hydrated_deployments( &mut self, request: impl IntoRequest<ListHydratedDeploymentsRequest> ) -> Result<Response<ListHydratedDeploymentsResponse>, Status>

List all hydrated deployments present under a deployment.

pub async fn update_hydrated_deployment( &mut self, request: impl IntoRequest<UpdateHydratedDeploymentRequest> ) -> Result<Response<HydratedDeployment>, Status>

Updates a hydrated deployment.

pub async fn apply_hydrated_deployment( &mut self, request: impl IntoRequest<ApplyHydratedDeploymentRequest> ) -> Result<Response<HydratedDeployment>, Status>

Applies a hydrated deployment to a workload cluster.