Module google_api_proto::google::cloud::texttospeech::v1beta1

Modules

custom_voice_params

Nested message and enum types in CustomVoiceParams.

streaming_synthesis_input

Nested message and enum types in StreamingSynthesisInput.

streaming_synthesize_request

Nested message and enum types in StreamingSynthesizeRequest.

synthesis_input

Nested message and enum types in SynthesisInput.

synthesize_speech_request

Nested message and enum types in SynthesizeSpeechRequest.

text_to_speech_client

Generated client implementations.

text_to_speech_long_audio_synthesize_client

Generated client implementations.

Structs

AudioConfig

Description of audio data to be synthesized.

CustomVoiceParams

Description of the custom voice to be synthesized.

ListVoicesRequest

The top-level message sent by the client for the ListVoices method.

ListVoicesResponse

The message returned to the client by the ListVoices method.

StreamingSynthesisInput

Input to be synthesized.

StreamingSynthesizeConfig

Provides configuration information for the StreamingSynthesize request.

StreamingSynthesizeRequest

Request message for the StreamingSynthesize method. Multiple StreamingSynthesizeRequest messages are sent in one call. The first message must contain a streaming_config that fully specifies the request configuration and must not contain input. All subsequent messages must only have input set.

StreamingSynthesizeResponse

StreamingSynthesizeResponse is the only message returned to the client by StreamingSynthesize method. A series of zero or more StreamingSynthesizeResponse messages are streamed back to the client.

SynthesisInput

Contains text input to be synthesized. Either text or ssml must be supplied. Supplying both or neither returns [google.rpc.Code.INVALID_ARGUMENT][google.rpc.Code.INVALID_ARGUMENT]. The input size is limited to 5000 bytes.

SynthesizeLongAudioMetadata

Metadata for response returned by the SynthesizeLongAudio method.

SynthesizeLongAudioRequest

The top-level message sent by the client for the SynthesizeLongAudio method.

SynthesizeLongAudioResponse

The message returned to the client by the SynthesizeLongAudio method.

SynthesizeSpeechRequest

The top-level message sent by the client for the SynthesizeSpeech method.

SynthesizeSpeechResponse

The message returned to the client by the SynthesizeSpeech method.

Timepoint

This contains a mapping between a certain point in the input text and a corresponding time in the output audio.

Voice

Description of a voice supported by the TTS service.

VoiceSelectionParams

Description of which voice to use for a synthesis request.

Enums

AudioEncoding

Configuration to set up audio encoder. The encoding determines the output audio format that we’d like.

SsmlVoiceGender

Gender of the voice as described in SSML voice element.