Module google_api_proto::google::cloud::video::stitcher::v1

Modules

cdn_key

Nested message and enum types in CdnKey.

companion

Nested message and enum types in Companion.

companion_ads

Nested message and enum types in CompanionAds.

event

Nested message and enum types in Event.

live_config

Nested message and enum types in LiveConfig.

live_session

Nested message and enum types in LiveSession.

manifest_options

Nested message and enum types in ManifestOptions.

media_cdn_key

Nested message and enum types in MediaCdnKey.

slate

Nested message and enum types in Slate.

video_stitcher_service_client

Generated client implementations.

vod_config

Nested message and enum types in VodConfig.

vod_session

Nested message and enum types in VodSession.

Structs

AdRequest

Details of an ad request to an ad server.

AdStitchDetail

Metadata for a stitched ad.

AkamaiCdnKey

Configuration for an Akamai CDN key.

CdnKey

Configuration for a CDN key. Used by the Video Stitcher to sign URIs for fetching video manifests and signing media segments for playback.

Companion

Metadata for a companion.

CompanionAds

Metadata for companion ads.

CreateCdnKeyRequest

Request message for VideoStitcherService.createCdnKey.

CreateLiveConfigRequest

Request message for VideoStitcherService.createLiveConfig

CreateLiveSessionRequest

Request message for VideoStitcherService.createLiveSession.

CreateSlateRequest

Request message for VideoStitcherService.createSlate.

CreateVodConfigRequest

Request message for VideoStitcherService.createVodConfig

CreateVodSessionRequest

Request message for VideoStitcherService.createVodSession

DeleteCdnKeyRequest

Request message for VideoStitcherService.deleteCdnKey.

DeleteLiveConfigRequest

Request message for VideoStitcherService.deleteLiveConfig.

DeleteSlateRequest

Request message for VideoStitcherService.deleteSlate.

DeleteVodConfigRequest

Request message for VideoStitcherService.deleteVodConfig.

Event

Describes an event and a trigger URI.

FetchOptions

Options on how fetches should be made.

GamLiveConfig

Metadata used to register a live stream with Google Ad Manager (GAM)

GamVodConfig

Metadata used for GAM ad decisioning.

GetCdnKeyRequest

Request message for VideoStitcherService.getCdnKey.

GetLiveAdTagDetailRequest

Request message for VideoStitcherService.getLiveAdTagDetail

GetLiveConfigRequest

Request message for VideoStitcherService.getLiveConfig.

GetLiveSessionRequest

Request message for VideoStitcherService.getSession.

GetSlateRequest

Request message for VideoStitcherService.getSlate.

GetVodAdTagDetailRequest

Request message for VideoStitcherService.getVodAdTagDetail

GetVodConfigRequest

Request message for VideoStitcherService.getVodConfig.

GetVodSessionRequest

Request message for VideoStitcherService.getVodSession

GetVodStitchDetailRequest

Request message for VideoStitcherService.getVodStitchDetail.

GoogleCdnKey

Configuration for a Google Cloud CDN key.

HtmlAdResource

Metadata for an HTML ad resource.

IframeAdResource

Metadata for an IFrame ad resource.

Interstitials

Describes what was stitched into a VOD session’s manifest.

ListCdnKeysRequest

Request message for VideoStitcherService.listCdnKeys.

ListCdnKeysResponse

Response message for VideoStitcher.ListCdnKeys.

ListLiveAdTagDetailsRequest

Request message for VideoStitcherService.listLiveAdTagDetails.

ListLiveAdTagDetailsResponse

Response message for VideoStitcherService.listLiveAdTagDetails.

ListLiveConfigsRequest

Request message for VideoStitcherService.listLiveConfig.

ListLiveConfigsResponse

Response message for VideoStitcher.ListLiveConfig.

ListSlatesRequest

Request message for VideoStitcherService.listSlates.

ListSlatesResponse

Response message for VideoStitcherService.listSlates.

ListVodAdTagDetailsRequest

Request message for VideoStitcherService.listVodAdTagDetails.

ListVodAdTagDetailsResponse

Response message for VideoStitcherService.listVodAdTagDetails.

ListVodConfigsRequest

Request message for VideoStitcherService.listVodConfig.

ListVodConfigsResponse

Response message for VideoStitcher.ListVodConfig.

ListVodStitchDetailsRequest

Request message for VideoStitcherService.listVodStitchDetails.

ListVodStitchDetailsResponse

Response message for VideoStitcherService.listVodStitchDetails.

LiveAdTagDetail

Information related to the details for one ad tag. This resource is only available for live sessions that do not implement Google Ad Manager ad insertion.

LiveConfig

Metadata for used to register live configs.

LiveSession

Metadata for a live session. The session expires 5 minutes after the client stops fetching the session’s playlists.

ManifestOptions

Options for manifest generation.

MediaCdnKey

Configuration for a Media CDN key.

OperationMetadata

Represents the metadata of the long-running operation.

PrefetchConfig

The configuration for prefetch ads.

ProgressEvent

Indicates a time in which a list of events should be triggered during media playback.

RenditionFilter

Filters for a video or muxed redition.

RequestMetadata

Metadata for an ad request.

ResponseMetadata

Metadata for the response of an ad request.

Slate

Slate object

StaticAdResource

Metadata for a static ad resource.

UpdateCdnKeyRequest

Request message for VideoStitcherService.updateCdnKey.

UpdateLiveConfigRequest

Request message for VideoStitcherService.updateLiveConfig.

UpdateSlateRequest

Request message for VideoStitcherService.updateSlate.

UpdateVodConfigRequest

Request message for VideoStitcherService.updateVodConfig.

VodAdTagDetail

Information related to the details for one ad tag. This resource is only available for VOD sessions that do not implement Google Ad Manager ad insertion.

VodConfig

Metadata used to register VOD configs.

VodSession

Metadata for a VOD session. The session expires 4 hours after its creation.

VodSessionAd

Metadata for an inserted ad in a VOD session.

VodSessionAdBreak

Metadata for an inserted ad break.

VodSessionContent

Metadata for the entire stitched content in a VOD session.

VodStitchDetail

Information related to the interstitial of a VOD session. This resource is only available for VOD sessions that do not implement Google Ad Manager ad insertion.

Enums

AdTracking

Determines the ad tracking policy.