Struct google_api_proto::google::cloud::video::stitcher::v1::video_stitcher_service_client::VideoStitcherServiceClient

pub struct VideoStitcherServiceClient<T> { /* private fields */ }

Video-On-Demand content stitching API allows you to insert ads into (VoD) video on demand files. You will be able to render custom scrubber bars with highlighted ads, enforce ad policies, allow seamless playback and tracking on native players and monetize content with any standard VMAP compliant ad server.

Implementations

impl<T> VideoStitcherServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> VideoStitcherServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_cdn_key( &mut self, request: impl IntoRequest<CreateCdnKeyRequest> ) -> Result<Response<Operation>, Status>

Creates a new CDN key.

pub async fn list_cdn_keys( &mut self, request: impl IntoRequest<ListCdnKeysRequest> ) -> Result<Response<ListCdnKeysResponse>, Status>

Lists all CDN keys in the specified project and location.

pub async fn get_cdn_key( &mut self, request: impl IntoRequest<GetCdnKeyRequest> ) -> Result<Response<CdnKey>, Status>

Returns the specified CDN key.

pub async fn delete_cdn_key( &mut self, request: impl IntoRequest<DeleteCdnKeyRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified CDN key.

pub async fn update_cdn_key( &mut self, request: impl IntoRequest<UpdateCdnKeyRequest> ) -> Result<Response<Operation>, Status>

Updates the specified CDN key. Only update fields specified in the call method body.

pub async fn create_vod_session( &mut self, request: impl IntoRequest<CreateVodSessionRequest> ) -> Result<Response<VodSession>, Status>

Creates a client side playback VOD session and returns the full tracking and playback metadata of the session.

pub async fn get_vod_session( &mut self, request: impl IntoRequest<GetVodSessionRequest> ) -> Result<Response<VodSession>, Status>

Returns the full tracking, playback metadata, and relevant ad-ops logs for the specified VOD session.

pub async fn list_vod_stitch_details( &mut self, request: impl IntoRequest<ListVodStitchDetailsRequest> ) -> Result<Response<ListVodStitchDetailsResponse>, Status>

Returns a list of detailed stitching information of the specified VOD session.

pub async fn get_vod_stitch_detail( &mut self, request: impl IntoRequest<GetVodStitchDetailRequest> ) -> Result<Response<VodStitchDetail>, Status>

Returns the specified stitching information for the specified VOD session.

pub async fn list_vod_ad_tag_details( &mut self, request: impl IntoRequest<ListVodAdTagDetailsRequest> ) -> Result<Response<ListVodAdTagDetailsResponse>, Status>

Return the list of ad tag details for the specified VOD session.

pub async fn get_vod_ad_tag_detail( &mut self, request: impl IntoRequest<GetVodAdTagDetailRequest> ) -> Result<Response<VodAdTagDetail>, Status>

Returns the specified ad tag detail for the specified VOD session.

pub async fn list_live_ad_tag_details( &mut self, request: impl IntoRequest<ListLiveAdTagDetailsRequest> ) -> Result<Response<ListLiveAdTagDetailsResponse>, Status>

Return the list of ad tag details for the specified live session.

pub async fn get_live_ad_tag_detail( &mut self, request: impl IntoRequest<GetLiveAdTagDetailRequest> ) -> Result<Response<LiveAdTagDetail>, Status>

Returns the specified ad tag detail for the specified live session.

pub async fn create_slate( &mut self, request: impl IntoRequest<CreateSlateRequest> ) -> Result<Response<Operation>, Status>

Creates a slate.

pub async fn list_slates( &mut self, request: impl IntoRequest<ListSlatesRequest> ) -> Result<Response<ListSlatesResponse>, Status>

Lists all slates in the specified project and location.

pub async fn get_slate( &mut self, request: impl IntoRequest<GetSlateRequest> ) -> Result<Response<Slate>, Status>

Returns the specified slate.

pub async fn update_slate( &mut self, request: impl IntoRequest<UpdateSlateRequest> ) -> Result<Response<Operation>, Status>

Updates the specified slate.

pub async fn delete_slate( &mut self, request: impl IntoRequest<DeleteSlateRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified slate.

pub async fn create_live_session( &mut self, request: impl IntoRequest<CreateLiveSessionRequest> ) -> Result<Response<LiveSession>, Status>

Creates a new live session.

pub async fn get_live_session( &mut self, request: impl IntoRequest<GetLiveSessionRequest> ) -> Result<Response<LiveSession>, Status>

Returns the details for the specified live session.

pub async fn create_live_config( &mut self, request: impl IntoRequest<CreateLiveConfigRequest> ) -> Result<Response<Operation>, Status>

Registers the live config with the provided unique ID in the specified region.

pub async fn list_live_configs( &mut self, request: impl IntoRequest<ListLiveConfigsRequest> ) -> Result<Response<ListLiveConfigsResponse>, Status>

Lists all live configs managed by the Video Stitcher that belong to the specified project and region.

pub async fn get_live_config( &mut self, request: impl IntoRequest<GetLiveConfigRequest> ) -> Result<Response<LiveConfig>, Status>

Returns the specified live config managed by the Video Stitcher service.

pub async fn delete_live_config( &mut self, request: impl IntoRequest<DeleteLiveConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified live config.

pub async fn update_live_config( &mut self, request: impl IntoRequest<UpdateLiveConfigRequest> ) -> Result<Response<Operation>, Status>

Updates the specified LiveConfig. Only update fields specified in the call method body.

pub async fn create_vod_config( &mut self, request: impl IntoRequest<CreateVodConfigRequest> ) -> Result<Response<Operation>, Status>

Registers the VOD config with the provided unique ID in the specified region.

pub async fn list_vod_configs( &mut self, request: impl IntoRequest<ListVodConfigsRequest> ) -> Result<Response<ListVodConfigsResponse>, Status>

Lists all VOD configs managed by the Video Stitcher API that belong to the specified project and region.

pub async fn get_vod_config( &mut self, request: impl IntoRequest<GetVodConfigRequest> ) -> Result<Response<VodConfig>, Status>

Returns the specified VOD config managed by the Video Stitcher API service.

pub async fn delete_vod_config( &mut self, request: impl IntoRequest<DeleteVodConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified VOD config.

pub async fn update_vod_config( &mut self, request: impl IntoRequest<UpdateVodConfigRequest> ) -> Result<Response<Operation>, Status>

Updates the specified VOD config. Only update fields specified in the call method body.