Module google_api_proto::google::cloud::video::transcoder::v1::overlay

Nested message and enum types in Overlay.

Modules

animation

Nested message and enum types in Animation.

Structs

Animation

Animation types.

AnimationEnd

End previous overlay animation from the video. Without AnimationEnd, the overlay object will keep the state of previous animation until the end of the video.

AnimationFade

Display overlay object with fade animation.

AnimationStatic

Display static overlay object.

Image

Overlaid image.

NormalizedCoordinate

2D normalized coordinates. Default: {0.0, 0.0}

Enums

FadeType

Fade type for the overlay: FADE_IN or FADE_OUT.