Module google_api_proto::google::cloud::videointelligence::v1beta2

Modules

video_intelligence_service_client

Generated client implementations.

Structs

AnnotateVideoProgress

Video annotation progress. Included in the metadata field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

AnnotateVideoRequest

Video annotation request.

AnnotateVideoResponse

Video annotation response. Included in the response field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

Entity

Detected entity from video analysis.

ExplicitContentAnnotation

Explicit content annotation (based on per-frame visual signals only). If no explicit content has been detected in a frame, no annotations are present for that frame.

ExplicitContentDetectionConfig

Config for EXPLICIT_CONTENT_DETECTION.

ExplicitContentFrame

Video frame level annotation results for explicit content.

FaceAnnotation

Face annotation.

FaceDetectionConfig

Config for FACE_DETECTION.

FaceFrame

Video frame level annotation results for face detection.

FaceSegment

Video segment level annotation results for face detection.

LabelAnnotation

Label annotation.

LabelDetectionConfig

Config for LABEL_DETECTION.

LabelFrame

Video frame level annotation results for label detection.

LabelSegment

Video segment level annotation results for label detection.

NormalizedBoundingBox

Normalized bounding box. The normalized vertex coordinates are relative to the original image. Range: [0, 1].

ShotChangeDetectionConfig

Config for SHOT_CHANGE_DETECTION.

VideoAnnotationProgress

Annotation progress for a single video.

VideoAnnotationResults

Annotation results for a single video.

VideoContext

Video context and/or feature-specific parameters.

VideoSegment

Video segment.

Enums

Feature

Video annotation feature.

LabelDetectionMode

Label detection mode.

Likelihood

Bucketized representation of likelihood.