Module google_api_proto::google::cloud::videointelligence::v1p1beta1

Modules

video_intelligence_service_client

Generated client implementations.

Structs

AnnotateVideoProgress

Video annotation progress. Included in the metadata field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

AnnotateVideoRequest

Video annotation request.

AnnotateVideoResponse

Video annotation response. Included in the response field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

Entity

Detected entity from video analysis.

ExplicitContentAnnotation

Explicit content annotation (based on per-frame visual signals only). If no explicit content has been detected in a frame, no annotations are present for that frame.

ExplicitContentDetectionConfig

Config for EXPLICIT_CONTENT_DETECTION.

ExplicitContentFrame

Video frame level annotation results for explicit content.

LabelAnnotation

Label annotation.

LabelDetectionConfig

Config for LABEL_DETECTION.

LabelFrame

Video frame level annotation results for label detection.

LabelSegment

Video segment level annotation results for label detection.

ShotChangeDetectionConfig

Config for SHOT_CHANGE_DETECTION.

SpeechContext

Provides “hints” to the speech recognizer to favor specific words and phrases in the results.

SpeechRecognitionAlternative

Alternative hypotheses (a.k.a. n-best list).

SpeechTranscription

A speech recognition result corresponding to a portion of the audio.

SpeechTranscriptionConfig

Config for SPEECH_TRANSCRIPTION.

VideoAnnotationProgress

Annotation progress for a single video.

VideoAnnotationResults

Annotation results for a single video.

VideoContext

Video context and/or feature-specific parameters.

VideoSegment

Video segment.

WordInfo

Word-specific information for recognized words. Word information is only included in the response when certain request parameters are set, such as enable_word_time_offsets.

Enums

Feature

Video annotation feature.

LabelDetectionMode

Label detection mode.

Likelihood

Bucketized representation of likelihood.