Module google_api_proto::google::cloud::videointelligence::v1p2beta1

Modules

object_tracking_annotation

Nested message and enum types in ObjectTrackingAnnotation.

video_intelligence_service_client

Generated client implementations.

Structs

AnnotateVideoProgress

Video annotation progress. Included in the metadata field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

AnnotateVideoRequest

Video annotation request.

AnnotateVideoResponse

Video annotation response. Included in the response field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

Entity

Detected entity from video analysis.

ExplicitContentAnnotation

Explicit content annotation (based on per-frame visual signals only). If no explicit content has been detected in a frame, no annotations are present for that frame.

ExplicitContentDetectionConfig

Config for EXPLICIT_CONTENT_DETECTION.

ExplicitContentFrame

Video frame level annotation results for explicit content.

LabelAnnotation

Label annotation.

LabelDetectionConfig

Config for LABEL_DETECTION.

LabelFrame

Video frame level annotation results for label detection.

LabelSegment

Video segment level annotation results for label detection.

NormalizedBoundingBox

Normalized bounding box. The normalized vertex coordinates are relative to the original image. Range: [0, 1].

NormalizedBoundingPoly

Normalized bounding polygon for text (that might not be aligned with axis). Contains list of the corner points in clockwise order starting from top-left corner. For example, for a rectangular bounding box: When the text is horizontal it might look like: 0––1 | | 3––2

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ObjectTrackingAnnotation

Annotations corresponding to one tracked object.

ObjectTrackingFrame

Video frame level annotations for object detection and tracking. This field stores per frame location, time offset, and confidence.

ShotChangeDetectionConfig

Config for SHOT_CHANGE_DETECTION.

TextAnnotation

Annotations related to one detected OCR text snippet. This will contain the corresponding text, confidence value, and frame level information for each detection.

TextDetectionConfig

Config for TEXT_DETECTION.

TextFrame

Video frame level annotation results for text annotation (OCR). Contains information regarding timestamp and bounding box locations for the frames containing detected OCR text snippets.

TextSegment

Video segment level annotation results for text detection.

VideoAnnotationProgress

Annotation progress for a single video.

VideoAnnotationResults

Annotation results for a single video.

VideoContext

Video context and/or feature-specific parameters.

VideoSegment

Video segment.

Enums

Feature

Video annotation feature.

LabelDetectionMode

Label detection mode.

Likelihood

Bucketized representation of likelihood.