Struct google_api_proto::google::cloud::vision::v1p1beta1::FaceAnnotation

pub struct FaceAnnotation {

    pub bounding_poly: Option<BoundingPoly>,
    pub fd_bounding_poly: Option<BoundingPoly>,
    pub landmarks: Vec<Landmark>,
    pub roll_angle: f32,
    pub pan_angle: f32,
    pub tilt_angle: f32,
    pub detection_confidence: f32,
    pub landmarking_confidence: f32,
    pub joy_likelihood: i32,
    pub sorrow_likelihood: i32,
    pub anger_likelihood: i32,
    pub surprise_likelihood: i32,
    pub under_exposed_likelihood: i32,
    pub blurred_likelihood: i32,
    pub headwear_likelihood: i32,

}

A face annotation object contains the results of face detection.

Fields
bounding_poly: Option<BoundingPoly>

The bounding polygon around the face. The coordinates of the bounding box are in the original image’s scale, as returned in ImageParams. The bounding box is computed to “frame” the face in accordance with human expectations. It is based on the landmarker results. Note that one or more x and/or y coordinates may not be generated in the BoundingPoly (the polygon will be unbounded) if only a partial face appears in the image to be annotated.

fd_bounding_poly: Option<BoundingPoly>

The fd_bounding_poly bounding polygon is tighter than the boundingPoly, and encloses only the skin part of the face. Typically, it is used to eliminate the face from any image analysis that detects the “amount of skin” visible in an image. It is not based on the landmarker results, only on the initial face detection, hence the fd (face detection) prefix.

landmarks: Vec<Landmark>

Detected face landmarks.

roll_angle: f32

Roll angle, which indicates the amount of clockwise/anti-clockwise rotation of the face relative to the image vertical about the axis perpendicular to the face. Range [-180,180].

pan_angle: f32

Yaw angle, which indicates the leftward/rightward angle that the face is pointing relative to the vertical plane perpendicular to the image. Range [-180,180].

tilt_angle: f32

Pitch angle, which indicates the upwards/downwards angle that the face is pointing relative to the image’s horizontal plane. Range [-180,180].

detection_confidence: f32

Detection confidence. Range [0, 1].

landmarking_confidence: f32

Face landmarking confidence. Range [0, 1].

joy_likelihood: i32

Joy likelihood.

sorrow_likelihood: i32

Sorrow likelihood.

anger_likelihood: i32

Anger likelihood.

surprise_likelihood: i32

Surprise likelihood.

under_exposed_likelihood: i32

Under-exposed likelihood.

blurred_likelihood: i32

Blurred likelihood.

headwear_likelihood: i32

Headwear likelihood.

Implementations

impl FaceAnnotation

pub fn joy_likelihood(&self) -> Likelihood

Returns the enum value of joy_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_joy_likelihood(&mut self, value: Likelihood)

Sets joy_likelihood to the provided enum value.

pub fn sorrow_likelihood(&self) -> Likelihood

Returns the enum value of sorrow_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_sorrow_likelihood(&mut self, value: Likelihood)

Sets sorrow_likelihood to the provided enum value.

pub fn anger_likelihood(&self) -> Likelihood

Returns the enum value of anger_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_anger_likelihood(&mut self, value: Likelihood)

Sets anger_likelihood to the provided enum value.

pub fn surprise_likelihood(&self) -> Likelihood

Returns the enum value of surprise_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_surprise_likelihood(&mut self, value: Likelihood)

Sets surprise_likelihood to the provided enum value.

pub fn under_exposed_likelihood(&self) -> Likelihood

Returns the enum value of under_exposed_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_under_exposed_likelihood(&mut self, value: Likelihood)

Sets under_exposed_likelihood to the provided enum value.

pub fn blurred_likelihood(&self) -> Likelihood

Returns the enum value of blurred_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_blurred_likelihood(&mut self, value: Likelihood)

Sets blurred_likelihood to the provided enum value.

pub fn headwear_likelihood(&self) -> Likelihood

Returns the enum value of headwear_likelihood, or the default if the field is set to an invalid enum value.

pub fn set_headwear_likelihood(&mut self, value: Likelihood)

Sets headwear_likelihood to the provided enum value.