Struct google_api_proto::google::cloud::visionai::v1::app_platform_client::AppPlatformClient

pub struct AppPlatformClient<T> { /* private fields */ }

Service describing handlers for resources

Implementations

impl<T> AppPlatformClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AppPlatformClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_applications( &mut self, request: impl IntoRequest<ListApplicationsRequest> ) -> Result<Response<ListApplicationsResponse>, Status>

Lists Applications in a given project and location.

pub async fn get_application( &mut self, request: impl IntoRequest<GetApplicationRequest> ) -> Result<Response<Application>, Status>

Gets details of a single Application.

pub async fn create_application( &mut self, request: impl IntoRequest<CreateApplicationRequest> ) -> Result<Response<Operation>, Status>

Creates a new Application in a given project and location.

pub async fn update_application( &mut self, request: impl IntoRequest<UpdateApplicationRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Application.

pub async fn delete_application( &mut self, request: impl IntoRequest<DeleteApplicationRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Application.

pub async fn deploy_application( &mut self, request: impl IntoRequest<DeployApplicationRequest> ) -> Result<Response<Operation>, Status>

Deploys a single Application.

pub async fn undeploy_application( &mut self, request: impl IntoRequest<UndeployApplicationRequest> ) -> Result<Response<Operation>, Status>

Undeploys a single Application.

pub async fn add_application_stream_input( &mut self, request: impl IntoRequest<AddApplicationStreamInputRequest> ) -> Result<Response<Operation>, Status>

Adds target stream input to the Application. If the Application is deployed, the corresponding new Application instance will be created. If the stream has already been in the Application, the RPC will fail.

pub async fn remove_application_stream_input( &mut self, request: impl IntoRequest<RemoveApplicationStreamInputRequest> ) -> Result<Response<Operation>, Status>

Remove target stream input to the Application, if the Application is deployed, the corresponding instance based will be deleted. If the stream is not in the Application, the RPC will fail.

pub async fn update_application_stream_input( &mut self, request: impl IntoRequest<UpdateApplicationStreamInputRequest> ) -> Result<Response<Operation>, Status>

Update target stream input to the Application, if the Application is deployed, the corresponding instance based will be deployed. For CreateOrUpdate behavior, set allow_missing to true.

pub async fn list_instances( &mut self, request: impl IntoRequest<ListInstancesRequest> ) -> Result<Response<ListInstancesResponse>, Status>

Lists Instances in a given project and location.

pub async fn get_instance( &mut self, request: impl IntoRequest<GetInstanceRequest> ) -> Result<Response<Instance>, Status>

Gets details of a single Instance.

pub async fn create_application_instances( &mut self, request: impl IntoRequest<CreateApplicationInstancesRequest> ) -> Result<Response<Operation>, Status>

Adds target stream input to the Application. If the Application is deployed, the corresponding new Application instance will be created. If the stream has already been in the Application, the RPC will fail.

pub async fn delete_application_instances( &mut self, request: impl IntoRequest<DeleteApplicationInstancesRequest> ) -> Result<Response<Operation>, Status>

Remove target stream input to the Application, if the Application is deployed, the corresponding instance based will be deleted. If the stream is not in the Application, the RPC will fail.

pub async fn update_application_instances( &mut self, request: impl IntoRequest<UpdateApplicationInstancesRequest> ) -> Result<Response<Operation>, Status>

Adds target stream input to the Application. If the Application is deployed, the corresponding new Application instance will be created. If the stream has already been in the Application, the RPC will fail.

pub async fn list_drafts( &mut self, request: impl IntoRequest<ListDraftsRequest> ) -> Result<Response<ListDraftsResponse>, Status>

Lists Drafts in a given project and location.

pub async fn get_draft( &mut self, request: impl IntoRequest<GetDraftRequest> ) -> Result<Response<Draft>, Status>

Gets details of a single Draft.

pub async fn create_draft( &mut self, request: impl IntoRequest<CreateDraftRequest> ) -> Result<Response<Operation>, Status>

Creates a new Draft in a given project and location.

pub async fn update_draft( &mut self, request: impl IntoRequest<UpdateDraftRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Draft.

pub async fn delete_draft( &mut self, request: impl IntoRequest<DeleteDraftRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Draft.

pub async fn list_processors( &mut self, request: impl IntoRequest<ListProcessorsRequest> ) -> Result<Response<ListProcessorsResponse>, Status>

Lists Processors in a given project and location.

pub async fn list_prebuilt_processors( &mut self, request: impl IntoRequest<ListPrebuiltProcessorsRequest> ) -> Result<Response<ListPrebuiltProcessorsResponse>, Status>

ListPrebuiltProcessors is a custom pass-through verb that Lists Prebuilt Processors.

pub async fn get_processor( &mut self, request: impl IntoRequest<GetProcessorRequest> ) -> Result<Response<Processor>, Status>

Gets details of a single Processor.

pub async fn create_processor( &mut self, request: impl IntoRequest<CreateProcessorRequest> ) -> Result<Response<Operation>, Status>

Creates a new Processor in a given project and location.

pub async fn update_processor( &mut self, request: impl IntoRequest<UpdateProcessorRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Processor.

pub async fn delete_processor( &mut self, request: impl IntoRequest<DeleteProcessorRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Processor.