Module google_api_proto::google::cloud::visionai::v1::video_object_tracking_prediction_result

Nested message and enum types in VideoObjectTrackingPredictionResult.

Structs

BoundingBox

Boundingbox for detected object. I.e. the rectangle over the video frame pinpointing the found AnnotationSpec. The coordinates are relative to the frame size, and the point 0,0 is in the top left of the frame.

DetectedObject

Each DetectedObject is one particular identification of an object specified with the AnnotationSpec id and display_name, the bounding box, the associated confidence score and the corresponding track_id.