Module google_api_proto::google::cloud::visionai::v1alpha1

Modules

analyzer_definition

Nested message and enum types in AnalyzerDefinition.

annotation_value

Nested message and enum types in AnnotationValue.

app_platform_client

Generated client implementations.

app_platform_cloud_function_request

Nested message and enum types in AppPlatformCloudFunctionRequest.

app_platform_cloud_function_response

Nested message and enum types in AppPlatformCloudFunctionResponse.

application

Nested message and enum types in Application.

application_configs

Nested message and enum types in ApplicationConfigs.

attribute_value

Nested message and enum types in AttributeValue.

clip_asset_response

Nested message and enum types in ClipAssetResponse.

cluster

Nested message and enum types in Cluster.

controlled_mode

Nested message and enum types in ControlledMode.

criteria

Nested message and enum types in Criteria.

custom_processor_source_info

Nested message and enum types in CustomProcessorSourceInfo.

data_schema_details

Nested message and enum types in DataSchemaDetails.

event

Nested message and enum types in Event.

facet_bucket

Nested message and enum types in FacetBucket.

facet_property

Nested message and enum types in FacetProperty.

facet_value

Nested message and enum types in FacetValue.

ingest_asset_request

Nested message and enum types in IngestAssetRequest.

instance

Nested message and enum types in Instance.

live_video_analytics_client

Generated client implementations.

node

Nested message and enum types in Node.

object_detection_prediction_result

Nested message and enum types in ObjectDetectionPredictionResult.

occupancy_counting_prediction_result

Nested message and enum types in OccupancyCountingPredictionResult.

packet_type

Nested message and enum types in PacketType.

partition

Nested message and enum types in Partition.

person_blur_config

Nested message and enum types in PersonBlurConfig.

personal_protective_equipment_detection_output

Nested message and enum types in PersonalProtectiveEquipmentDetectionOutput.

processor

Nested message and enum types in Processor.

processor_config

Nested message and enum types in ProcessorConfig.

processor_io_spec

Nested message and enum types in ProcessorIOSpec.

receive_events_control_response

Nested message and enum types in ReceiveEventsControlResponse.

receive_events_request

Nested message and enum types in ReceiveEventsRequest.

receive_events_response

Nested message and enum types in ReceiveEventsResponse.

receive_packets_control_response

Nested message and enum types in ReceivePacketsControlResponse.

receive_packets_request

Nested message and enum types in ReceivePacketsRequest.

receive_packets_response

Nested message and enum types in ReceivePacketsResponse.

remove_application_stream_input_request

Nested message and enum types in RemoveApplicationStreamInputRequest.

send_packets_request

Nested message and enum types in SendPacketsRequest.

stream_annotation

Nested message and enum types in StreamAnnotation.

stream_with_annotation

Nested message and enum types in StreamWithAnnotation.

streaming_service_client

Generated client implementations.

streams_service_client

Generated client implementations.

update_application_instances_request

Nested message and enum types in UpdateApplicationInstancesRequest.

video_action_recognition_prediction_result

Nested message and enum types in VideoActionRecognitionPredictionResult.

video_classification_prediction_result

Nested message and enum types in VideoClassificationPredictionResult.

video_object_tracking_prediction_result

Nested message and enum types in VideoObjectTrackingPredictionResult.

warehouse_client

Generated client implementations.

Structs

AcquireLeaseRequest

Request message for acquiring a lease.

AddApplicationStreamInputRequest

Message for adding stream input to an Application.

AddApplicationStreamInputResponse

Message for AddApplicationStreamInput Response.

AiEnabledDevicesInputConfig

Message describing AI-enabled Devices Input Config.

Analysis

Message describing the Analysis object.

AnalysisDefinition

Defines a full analysis.

AnalyzerDefinition

Defines an Analyzer.

Annotation

An annotation is a resource in asset. It represents a key-value mapping of content in asset.

AnnotationMatchingResult

Stores the criteria-annotation matching results for each search result item.

AnnotationValue

Value of annotation, including all types available in data schema.

AppPlatformCloudFunctionRequest

For any cloud function based customer processing logic, customer’s cloud function is expected to receive AppPlatformCloudFunctionRequest as request and send back AppPlatformCloudFunctionResponse as response. Message of request from AppPlatform to Cloud Function.

AppPlatformCloudFunctionResponse

Message of the response from customer’s Cloud Function to AppPlatform.

AppPlatformEventBody

Message of content of appPlatform event

AppPlatformMetadata

Message of essential metadata of App Platform. This message is usually attached to a certain processor output annotation for customer to identify the source of the data.

Application

Message describing Application object

ApplicationConfigs

Message storing the graph of the application.

ApplicationInstance

Message for creating a Instance.

ApplicationNodeAnnotation

Message describing annotations specific to application node. This message is a duplication of StreamWithAnnotation.NodeAnnotation.

ApplicationStreamInput

Message about a single stream input config.

Asset

An asset is a resource in corpus. It represents a media object inside corpus, contains metadata and another resource annotation. Different feature could be applied to the asset to generate annotations. User could specified annotation related to the target asset.

AttributeValue

Represents an actual value of an operator attribute.

AutoscalingMetricSpec

The metric specification that defines the target resource utilization (CPU utilization, accelerator’s duty cycle, and so on) for calculating the desired replica count.

BigQueryConfig

Message of configurations for BigQuery processor.

BoolValue

Channel

Message describing the Channel object.

CircleArea

Representation of a circle area.

ClassificationPredictionResult

Prediction output format for Image and Text Classification.

ClipAssetRequest

Request message for ClipAsset API.

ClipAssetResponse

Response message for ClipAsset API.

Cluster

Message describing the Cluster object.

CommitRequest

The message for explicitly committing the read progress.

ControlledMode

The options for receiver under the controlled mode.

Corpus

Corpus is a set of video contents for management. Within a corpus, videos share the same data schema. Search is also restricted within a single corpus.

CreateAnalysisRequest

Message for creating an Analysis.

CreateAnnotationRequest

Request message for CreateAnnotation.

CreateApplicationInstancesRequest

Message for adding stream input to an Application.

CreateApplicationInstancesResponse

Message for CreateApplicationInstance Response.

CreateApplicationRequest

Message for creating a Application.

CreateAssetRequest

Request message for CreateAssetRequest.

CreateClusterRequest

Message for creating a Cluster.

CreateCorpusMetadata

Metadata for CreateCorpus API.

CreateCorpusRequest

Request message of CreateCorpus API.

CreateDataSchemaRequest

Request message for CreateDataSchema.

CreateDraftRequest

Message for creating a Draft.

CreateEventRequest

Message for creating a Event.

CreateProcessorRequest

Message for creating a Processor.

CreateSearchConfigRequest

Request message for CreateSearchConfig.

CreateSeriesRequest

Message for creating a Series.

CreateStreamRequest

Message for creating a Stream.

Criteria

Filter criteria applied to current search results.

CustomProcessorSourceInfo

Describes the source info for a custom processor.

DataSchema

Data schema indicates how the user specified annotation is interpreted in the system.

DataSchemaDetails

Data schema details indicates the data type and the data struct corresponding to the key of user specified annotation.

DateTimeRange

Datetime range type.

DateTimeRangeArray

A list of datetime range values.

DedicatedResources

A description of resources that are dedicated to a DeployedModel, and that need a higher degree of manual configuration.

DeleteAnalysisRequest

Message for deleting an Analysis.

DeleteAnnotationRequest

Request message for DeleteAnnotation API.

DeleteApplicationInstancesRequest

Message for removing stream input from an Application.

DeleteApplicationInstancesResponse

Message for DeleteApplicationInstance Response.

DeleteApplicationRequest

Message for deleting an Application.

DeleteAssetMetadata

The metadata for DeleteAsset API that embeds in [metadata][google.longrunning.Operation.metadata] field.

DeleteAssetRequest

Request message for DeleteAsset.

DeleteClusterRequest

Message for deleting a Cluster.

DeleteCorpusRequest

Request message for DeleteCorpus.

DeleteDataSchemaRequest

Request message for DeleteDataSchema.

DeleteDraftRequest

Message for deleting an Draft.

DeleteEventRequest

Message for deleting a Event.

DeleteProcessorRequest

Message for deleting a Processor.

DeleteSearchConfigRequest

Request message for DeleteSearchConfig.

DeleteSeriesRequest

Message for deleting a Series.

DeleteStreamRequest

Message for deleting a Stream.

DeployApplicationRequest

Message for deploying an Application.

DeployApplicationResponse

RPC Request Messages. Message for DeployApplication Response.

Draft

Message describing Draft object

EagerMode

The options for receiver under the eager mode.

Event

Message describing the Event object.

EventUpdate

The event update message.

FacetBucket

Holds the facet value, selections state, and metadata.

FacetGroup

A group of facet buckets to be passed back and forth between backend & frontend.

FacetProperty

Central configuration for a facet.

FacetValue

Definition of a single value with generic type.

FloatRange

Float range type.

FloatRangeArray

A list of float range values.

GcsSource

The Google Cloud Storage location for the input content.

GeneralObjectDetectionConfig

Message of configurations for General Object Detection processor.

GenerateHlsUriRequest

Request message for GenerateHlsUri API.

GenerateHlsUriResponse

Response message for GenerateHlsUri API.

GenerateStreamHlsTokenRequest

Request message for getting the auth token to access the stream HLS contents.

GenerateStreamHlsTokenResponse

Response message for GenerateStreamHlsToken.

GeoCoordinate

Location Coordinate Representation

GeoLocationArray

A list of locations.

GetAnalysisRequest

Message for getting an Analysis.

GetAnnotationRequest

Request message for GetAnnotation API.

GetApplicationRequest

Message for getting a Application.

GetAssetRequest

Request message for GetAsset.

GetClusterRequest

Message for getting a Cluster.

GetCorpusRequest

Request message for GetCorpus.

GetDataSchemaRequest

Request message for GetDataSchema.

GetDraftRequest

Message for getting a Draft.

GetEventRequest

Message for getting a Event.

GetInstanceRequest

Message for getting a Instance.

GetProcessorRequest

Message for getting a Processor.

GetSearchConfigRequest

Request message for GetSearchConfig.

GetSeriesRequest

Message for getting a Series.

GetStreamRequest

Message for getting a Stream.

GetStreamThumbnailResponse

Message for the response of GetStreamThumbnail. The empty response message indicates the thumbnail image has been uploaded to GCS successfully.

GstreamerBufferDescriptor

The descriptor for a gstreamer buffer payload.

ImageObjectDetectionPredictionResult

Prediction output format for Image Object Detection.

ImageSegmentationPredictionResult

Prediction output format for Image Segmentation.

IngestAssetRequest

Request message for IngestAsset API.

IngestAssetResponse

Response message for IngestAsset API.

Instance

Message describing Instance object

IntRange

Integer range type.

IntRangeArray

A list of integer range values.

Lease

The lease message.

ListAnalysesRequest

Message for requesting list of Analyses

ListAnalysesResponse

Message for response to listing Analyses

ListAnnotationsRequest

Request message for GetAnnotation API.

ListAnnotationsResponse

Request message for ListAnnotations API.

ListApplicationsRequest

Message for requesting list of Applications.

ListApplicationsResponse

Message for response to listing Applications.

ListAssetsRequest

Request message for ListAssets.

ListAssetsResponse

Response message for ListAssets.

ListClustersRequest

Message for requesting list of Clusters.

ListClustersResponse

Message for response to listing Clusters.

ListCorporaRequest

Request message for ListCorpora.

ListCorporaResponse

Response message for ListCorpora.

ListDataSchemasRequest

Request message for ListDataSchemas.

ListDataSchemasResponse

Response message for ListDataSchemas.

ListDraftsRequest

Message for requesting list of Drafts.

ListDraftsResponse

Message for response to listing Drafts.

ListEventsRequest

Message for requesting list of Events.

ListEventsResponse

Message for response to listing Events.

ListInstancesRequest

Message for requesting list of Instances.

ListInstancesResponse

Message for response to listing Instances.

ListPrebuiltProcessorsRequest

Request Message for listing Prebuilt Processors.

ListPrebuiltProcessorsResponse

Response Message for listing Prebuilt Processors.

ListProcessorsRequest

Message for requesting list of Processors.

ListProcessorsResponse

Message for response to listing Processors.

ListSearchConfigsRequest

Request message for ListSearchConfigs.

ListSearchConfigsResponse

Response message for ListSearchConfigs.

ListSeriesRequest

Message for requesting list of Series.

ListSeriesResponse

Message for response to listing Series.

ListStreamsRequest

Message for requesting list of Streams.

ListStreamsResponse

Message for response to listing Streams.

MachineSpec

Specification of a single machine.

MaterializeChannelRequest

Message for materializing a channel.

MediaWarehouseConfig

Message describing MediaWarehouseConfig.

Node

Message describing node object.

NormalizedPolygon

Normalized Polygon.

NormalizedPolyline

Normalized Pplyline, which represents a curve consisting of connected straight-line segments.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ObjectDetectionPredictionResult

Prediction output format for Generic Object Detection.

OccupancyCountConfig

Message describing OccupancyCountConfig.

OccupancyCountingPredictionResult

The prediction result proto for occupancy counting.

OperationMetadata

Represents the metadata of the long-running operation.

Packet

The quanta of datum that the series accepts.

PacketHeader

The message that represents packet header.

PacketType

The message that represents the data type of a packet.

Partition

Partition to specify the partition in time and space for sub-asset level annotation.

PersonBlurConfig

Message describing FaceBlurConfig.

PersonVehicleDetectionConfig

Message describing PersonVehicleDetectionConfig.

PersonalProtectiveEquipmentDetectionConfig

Message describing PersonalProtectiveEquipmentDetectionConfig.

PersonalProtectiveEquipmentDetectionOutput

Output format for Personal Protective Equipment Detection Operator.

Processor

Message describing Processor object. Next ID: 18

ProcessorConfig

Next ID: 24

ProcessorIoSpec

Message describing the input / output specifications of a processor.

RawImageDescriptor

The descriptor for a raw image.

ReceiveEventsControlResponse

Control message for a ReceiveEventsResponse.

ReceiveEventsRequest

Request message for ReceiveEvents.

ReceiveEventsResponse

Response message for the ReceiveEvents.

ReceivePacketsControlResponse

Control message for a ReceivePacketsResponse.

ReceivePacketsRequest

Request message for receiving packets.

ReceivePacketsResponse

Response message from ReceivePackets.

ReleaseLeaseRequest

Request message for releasing lease.

ReleaseLeaseResponse

Response message for release lease.

RemoveApplicationStreamInputRequest

Message for removing stream input from an Application.

RemoveApplicationStreamInputResponse

Message for RemoveApplicationStreamInput Response.

RenewLeaseRequest

Request message for renewing a lease.

RequestMetadata

RequestMetadata is the metadata message for the request.

ResourceAnnotations

Message describing general annotation for resources.

SearchAssetsRequest

Request message for SearchAssets.

SearchAssetsResponse

Response message for SearchAssets.

SearchConfig

SearchConfig stores different properties that will affect search behaviors and search results.

SearchCriteriaProperty

Central configuration for custom search criteria.

SearchResultItem

Search result contains asset name and corresponding time ranges.

SendPacketsRequest

Request message for sending packets.

SendPacketsResponse

Response message for sending packets.

Series

Message describing the Series object.

SeriesMetadata

The message that represents series metadata.

ServerMetadata

The message that represents server metadata.

Stream

Message describing the Stream object. The Stream and the Event resources are many to many; i.e., each Stream resource can associate to many Event resources and each Event resource can associate to many Stream resources.

StreamAnnotation

message about annotations about Vision AI stream resource.

StreamAnnotations

A wrapper of repeated StreamAnnotation.

StreamWithAnnotation

Message describing Vision AI stream with application specific annotations. All the StreamAnnotation object inside this message MUST have unique id.

StringArray

A list of string-type values.

UndeployApplicationRequest

Message for undeploying an Application.

UndeployApplicationResponse

Message for UndeployApplication Response.

UpdateAnalysisRequest

Message for updating an Analysis.

UpdateAnnotationRequest

Request message for UpdateAnnotation API.

UpdateApplicationInstancesRequest

Message for updating an ApplicationInstance.

UpdateApplicationInstancesResponse

Message for UpdateApplicationInstances Response.

UpdateApplicationRequest

Message for updating an Application.

UpdateApplicationStreamInputRequest

Message for updating stream input to an Application.

UpdateApplicationStreamInputResponse

Message for AddApplicationStreamInput Response.

UpdateAssetRequest

Response message for UpdateAsset.

UpdateClusterRequest

Message for updating a Cluster.

UpdateCorpusRequest

Request message for UpdateCorpus.

UpdateDataSchemaRequest

Request message for UpdateDataSchema.

UpdateDraftRequest

Message for updating an Draft.

UpdateEventRequest

Message for updating a Event.

UpdateProcessorRequest

Message for updating a Processor.

UpdateSearchConfigRequest

Request message for UpdateSearchConfig.

UpdateSeriesRequest

Message for updating a Series.

UpdateStreamRequest

Message for updating a Stream.

UserSpecifiedAnnotation

Annotation provided by users.

VertexAutoMlVideoConfig

Message describing VertexAutoMLVideoConfig.

VertexAutoMlVisionConfig

Message of configurations of Vertex AutoML Vision Processors.

VertexCustomConfig

Message describing VertexCustomConfig.

VideoActionRecognitionPredictionResult

Prediction output format for Video Action Recognition.

VideoClassificationPredictionResult

Prediction output format for Video Classification.

VideoObjectTrackingPredictionResult

Prediction output format for Video Object Tracking.

VideoStreamInputConfig

Message describing Video Stream Input Config. This message should only be used as a placeholder for builtin:stream-input processor, actual stream binding should be specified using corresponding API.

Enums

AcceleratorType

Represents a hardware accelerator type.

FacetBucketType

Different types for a facet bucket.

LeaseType

The lease type.

ModelType

All the supported model types in Vision AI App Platform.

StreamAnnotationType

Enum describing all possible types of a stream annotation.