Module google_api_proto::google::cloud::vmmigration::v1

Modules

applied_license

Nested message and enum types in AppliedLicense.

aws_source_details

Nested message and enum types in AwsSourceDetails.

aws_source_vm_details

Nested message and enum types in AwsSourceVmDetails.

aws_vm_details

Nested message and enum types in AwsVmDetails.

clone_job

Nested message and enum types in CloneJob.

clone_step

Nested message and enum types in CloneStep.

compute_scheduling

Nested message and enum types in ComputeScheduling.

cutover_job

Nested message and enum types in CutoverJob.

cutover_step

Nested message and enum types in CutoverStep.

cycle_step

Nested message and enum types in CycleStep.

datacenter_connector

Nested message and enum types in DatacenterConnector.

fetch_inventory_response

Nested message and enum types in FetchInventoryResponse.

migrating_vm

Nested message and enum types in MigratingVm.

migration_error

Nested message and enum types in MigrationError.

replication_cycle

Nested message and enum types in ReplicationCycle.

scheduling_node_affinity

Nested message and enum types in SchedulingNodeAffinity.

source

Nested message and enum types in Source.

upgrade_status

Nested message and enum types in UpgradeStatus.

utilization_report

Nested message and enum types in UtilizationReport.

vm_migration_client

Generated client implementations.

vm_utilization_info

Nested message and enum types in VmUtilizationInfo.

vmware_vm_details

Nested message and enum types in VmwareVmDetails.

Structs

AdaptingOsStep

AdaptingOSStep contains specific step details.

AddGroupMigrationRequest

Request message for ‘AddGroupMigration’ request.

AddGroupMigrationResponse

Response message for ‘AddGroupMigration’ request.

ApplianceVersion

Describes an appliance version.

AppliedLicense

AppliedLicense holds the license data returned by adaptation module report.

AvailableUpdates

Holds informatiom about the available versions for upgrade.

AwsSecurityGroup

AwsSecurityGroup describes a security group of an AWS VM.

AwsSourceDetails

AwsSourceDetails message describes a specific source details for the AWS source type.

AwsSourceVmDetails

Represent the source AWS VM details.

AwsVmDetails

AwsVmDetails describes a VM in AWS.

AwsVmsDetails

AWSVmsDetails describes VMs in AWS.

CancelCloneJobRequest

Request message for ‘CancelCloneJob’ request.

CancelCloneJobResponse

Response message for ‘CancelCloneJob’ request.

CancelCutoverJobRequest

Request message for ‘CancelCutoverJob’ request.

CancelCutoverJobResponse

Response message for ‘CancelCutoverJob’ request.

CloneJob

CloneJob describes the process of creating a clone of a [MigratingVM][google.cloud.vmmigration.v1.MigratingVm] to the requested target based on the latest successful uploaded snapshots. While the migration cycles of a MigratingVm take place, it is possible to verify the uploaded VM can be started in the cloud, by creating a clone. The clone can be created without any downtime, and it is created using the latest snapshots which are already in the cloud. The cloneJob is only responsible for its work, not its products, which means once it is finished, it will never touch the instance it created. It will only delete it in case of the CloneJob being cancelled or upon failure to clone.

CloneStep

CloneStep holds information about the clone step progress.

ComputeEngineTargetDefaults

ComputeEngineTargetDefaults is a collection of details for creating a VM in a target Compute Engine project.

ComputeEngineTargetDetails

ComputeEngineTargetDetails is a collection of details for creating a VM in a target Compute Engine project.

ComputeScheduling

Scheduling information for VM on maintenance/restart behaviour and node allocation in sole tenant nodes.

CreateCloneJobRequest

Request message for ‘CreateCloneJob’ request.

CreateCutoverJobRequest

Request message for ‘CreateCutoverJob’ request.

CreateDatacenterConnectorRequest

Request message for ‘CreateDatacenterConnector’ request.

CreateGroupRequest

Request message for ‘CreateGroup’ request.

CreateMigratingVmRequest

Request message for ‘CreateMigratingVm’ request.

CreateSourceRequest

Request message for ‘CreateSource’ request.

CreateTargetProjectRequest

Request message for ‘CreateTargetProject’ request.

CreateUtilizationReportRequest

Request message for ‘CreateUtilizationReport’ request.

CutoverJob

CutoverJob message describes a cutover of a migrating VM. The CutoverJob is the operation of shutting down the VM, creating a snapshot and clonning the VM using the replicated snapshot.

CutoverStep

CutoverStep holds information about the cutover step progress.

CycleStep

CycleStep holds information about a step progress.

DatacenterConnector

DatacenterConnector message describes a connector between the Source and Google Cloud, which is installed on a vmware datacenter (an OVA vm installed by the user) to connect the Datacenter to Google Cloud and support vm migration data transfer.

DeleteDatacenterConnectorRequest

Request message for ‘DeleteDatacenterConnector’ request.

DeleteGroupRequest

Request message for ‘DeleteGroup’ request.

DeleteMigratingVmRequest

Request message for ‘DeleteMigratingVm’ request.

DeleteSourceRequest

Request message for ‘DeleteSource’ request.

DeleteTargetProjectRequest

Request message for ‘DeleteTargetProject’ request.

DeleteUtilizationReportRequest

Request message for ‘DeleteUtilizationReport’ request.

FetchInventoryRequest

Request message for [fetchInventory][google.cloud.vmmigration.v1.VmMigration.FetchInventory].

FetchInventoryResponse

Response message for [fetchInventory][google.cloud.vmmigration.v1.VmMigration.FetchInventory].

FinalizeMigrationRequest

Request message for ‘FinalizeMigration’ request.

FinalizeMigrationResponse

Response message for ‘FinalizeMigration’ request.

GetCloneJobRequest

Request message for ‘GetCloneJob’ request.

GetCutoverJobRequest

Request message for ‘GetCutoverJob’ request.

GetDatacenterConnectorRequest

Request message for ‘GetDatacenterConnector’ request.

GetGroupRequest

Request message for ‘GetGroup’ request.

GetMigratingVmRequest

Request message for ‘GetMigratingVm’ request.

GetReplicationCycleRequest

Request message for ‘GetReplicationCycle’ request.

GetSourceRequest

Request message for ‘GetSource’ request.

GetTargetProjectRequest

Request message for ‘GetTargetProject’ call.

GetUtilizationReportRequest

Request message for ‘GetUtilizationReport’ request.

Group

Describes message for ‘Group’ resource. The Group is a collections of several MigratingVms.

InitializingReplicationStep

InitializingReplicationStep contains specific step details.

InstantiatingMigratedVmStep

InstantiatingMigratedVMStep contains specific step details.

ListCloneJobsRequest

Request message for ‘ListCloneJobsRequest’ request.

ListCloneJobsResponse

Response message for ‘ListCloneJobs’ request.

ListCutoverJobsRequest

Request message for ‘ListCutoverJobsRequest’ request.

ListCutoverJobsResponse

Response message for ‘ListCutoverJobs’ request.

ListDatacenterConnectorsRequest

Request message for ‘ListDatacenterConnectors’ request.

ListDatacenterConnectorsResponse

Response message for ‘ListDatacenterConnectors’ request.

ListGroupsRequest

Request message for ‘ListGroups’ request.

ListGroupsResponse

Response message for ‘ListGroups’ request.

ListMigratingVmsRequest

Request message for ‘LisMigratingVmsRequest’ request.

ListMigratingVmsResponse

Response message for ‘ListMigratingVms’ request.

ListReplicationCyclesRequest

Request message for ‘LisReplicationCyclesRequest’ request.

ListReplicationCyclesResponse

Response message for ‘ListReplicationCycles’ request.

ListSourcesRequest

Request message for ‘ListSources’ request.

ListSourcesResponse

Response message for ‘ListSources’ request.

ListTargetProjectsRequest

Request message for ‘ListTargetProjects’ call.

ListTargetProjectsResponse

Response message for ‘ListTargetProjects’ call.

ListUtilizationReportsRequest

Request message for ‘ListUtilizationReports’ request.

ListUtilizationReportsResponse

Response message for ‘ListUtilizationReports’ request.

MigratingVm

MigratingVm describes the VM that will be migrated from a Source environment and its replication state.

MigrationError

Represents migration resource error information that can be used with google.rpc.Status message. MigrationError is used to present the user with error information in migration operations.

NetworkInterface

NetworkInterface represents a NIC of a VM.

OperationMetadata

Represents the metadata of the long-running operation.

PauseMigrationRequest

Request message for ‘PauseMigration’ request.

PauseMigrationResponse

Response message for ‘PauseMigration’ request.

PostProcessingStep

PostProcessingStep contains specific step details.

PreparingVmDisksStep

PreparingVMDisksStep contains specific step details.

RemoveGroupMigrationRequest

Request message for ‘RemoveMigration’ request.

RemoveGroupMigrationResponse

Response message for ‘RemoveMigration’ request.

ReplicatingStep

ReplicatingStep contains specific step details.

ReplicationCycle

ReplicationCycle contains information about the current replication cycle status.

ReplicationSync

ReplicationSync contain information about the last replica sync to the cloud.

ResumeMigrationRequest

Request message for ‘ResumeMigration’ request.

ResumeMigrationResponse

Response message for ‘ResumeMigration’ request.

SchedulePolicy

A policy for scheduling replications.

SchedulingNodeAffinity

Node Affinity: the configuration of desired nodes onto which this Instance could be scheduled. Based on https://cloud.google.com/compute/docs/reference/rest/v1/instances/setScheduling

ShuttingDownSourceVmStep

ShuttingDownSourceVMStep contains specific step details.

Source

Source message describes a specific vm migration Source resource. It contains the source environment information.

StartMigrationRequest

Request message for ‘StartMigrationRequest’ request.

StartMigrationResponse

Response message for ‘StartMigration’ request.

TargetProject

TargetProject message represents a target Compute Engine project for a migration or a clone.

UpdateGroupRequest

Update message for ‘UpdateGroups’ request.

UpdateMigratingVmRequest

Request message for ‘UpdateMigratingVm’ request.

UpdateSourceRequest

Update message for ‘UpdateSources’ request.

UpdateTargetProjectRequest

Update message for ‘UpdateTargetProject’ request.

UpgradeApplianceRequest

Request message for ‘UpgradeAppliance’ request.

UpgradeApplianceResponse

Response message for ‘UpgradeAppliance’ request.

UpgradeStatus

UpgradeStatus contains information about upgradeAppliance operation.

UtilizationReport

Utilization report details the utilization (CPU, memory, etc.) of selected source VMs.

VmUtilizationInfo

Utilization information of a single VM.

VmUtilizationMetrics

Utilization metrics values for a single VM.

VmwareSourceDetails

VmwareSourceDetails message describes a specific source details for the vmware source type.

VmwareVmDetails

VmwareVmDetails describes a VM in vCenter.

VmwareVmsDetails

VmwareVmsDetails describes VMs in vCenter.

Enums

ComputeEngineBootOption

Possible values for vm boot option.

ComputeEngineDiskType

Types of disks supported for Compute Engine VM.

ComputeEngineLicenseType

Types of licenses used in OS adaptation.

MigratingVmView

Controls the level of details of a Migrating VM.

UtilizationReportView

Controls the level of details of a Utilization Report.