Struct google_api_proto::google::cloud::vmwareengine::v1::vmware_engine_client::VmwareEngineClient

pub struct VmwareEngineClient<T> { /* private fields */ }

VMwareEngine manages VMware’s private clusters in the Cloud.

Implementations

impl<T> VmwareEngineClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> VmwareEngineClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_private_clouds( &mut self, request: impl IntoRequest<ListPrivateCloudsRequest>, ) -> Result<Response<ListPrivateCloudsResponse>, Status>

Lists PrivateCloud resources in a given project and location.

pub async fn get_private_cloud( &mut self, request: impl IntoRequest<GetPrivateCloudRequest>, ) -> Result<Response<PrivateCloud>, Status>

Retrieves a PrivateCloud resource by its resource name.

pub async fn create_private_cloud( &mut self, request: impl IntoRequest<CreatePrivateCloudRequest>, ) -> Result<Response<Operation>, Status>

Creates a new PrivateCloud resource in a given project and location. Private clouds of type STANDARD and TIME_LIMITED are zonal resources, STRETCHED private clouds are regional. Creating a private cloud also creates a management cluster for that private cloud.

pub async fn update_private_cloud( &mut self, request: impl IntoRequest<UpdatePrivateCloudRequest>, ) -> Result<Response<Operation>, Status>

Modifies a PrivateCloud resource. Only the following fields can be updated: description. Only fields specified in updateMask are applied.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_private_cloud( &mut self, request: impl IntoRequest<DeletePrivateCloudRequest>, ) -> Result<Response<Operation>, Status>

Schedules a PrivateCloud resource for deletion.

A PrivateCloud resource scheduled for deletion has PrivateCloud.state set to DELETED and expireTime set to the time when deletion is final and can no longer be reversed. The delete operation is marked as done as soon as the PrivateCloud is successfully scheduled for deletion (this also applies when delayHours is set to zero), and the operation is not kept in pending state until PrivateCloud is purged. PrivateCloud can be restored using UndeletePrivateCloud method before the expireTime elapses. When expireTime is reached, deletion is final and all private cloud resources are irreversibly removed and billing stops. During the final removal process, PrivateCloud.state is set to PURGING. PrivateCloud can be polled using standard GET method for the whole period of deletion and purging. It will not be returned only when it is completely purged.

pub async fn undelete_private_cloud( &mut self, request: impl IntoRequest<UndeletePrivateCloudRequest>, ) -> Result<Response<Operation>, Status>

Restores a private cloud that was previously scheduled for deletion by DeletePrivateCloud. A PrivateCloud resource scheduled for deletion has PrivateCloud.state set to DELETED and PrivateCloud.expireTime set to the time when deletion can no longer be reversed.

pub async fn list_clusters( &mut self, request: impl IntoRequest<ListClustersRequest>, ) -> Result<Response<ListClustersResponse>, Status>

Lists Cluster resources in a given private cloud.

pub async fn get_cluster( &mut self, request: impl IntoRequest<GetClusterRequest>, ) -> Result<Response<Cluster>, Status>

Retrieves a Cluster resource by its resource name.

pub async fn create_cluster( &mut self, request: impl IntoRequest<CreateClusterRequest>, ) -> Result<Response<Operation>, Status>

Creates a new cluster in a given private cloud. Creating a new cluster provides additional nodes for use in the parent private cloud and requires sufficient node quota.

pub async fn update_cluster( &mut self, request: impl IntoRequest<UpdateClusterRequest>, ) -> Result<Response<Operation>, Status>

Modifies a Cluster resource. Only fields specified in updateMask are applied.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_cluster( &mut self, request: impl IntoRequest<DeleteClusterRequest>, ) -> Result<Response<Operation>, Status>

Deletes a Cluster resource. To avoid unintended data loss, migrate or gracefully shut down any workloads running on the cluster before deletion. You cannot delete the management cluster of a private cloud using this method.

pub async fn list_nodes( &mut self, request: impl IntoRequest<ListNodesRequest>, ) -> Result<Response<ListNodesResponse>, Status>

Lists nodes in a given cluster.

pub async fn get_node( &mut self, request: impl IntoRequest<GetNodeRequest>, ) -> Result<Response<Node>, Status>

Gets details of a single node.

pub async fn list_external_addresses( &mut self, request: impl IntoRequest<ListExternalAddressesRequest>, ) -> Result<Response<ListExternalAddressesResponse>, Status>

Lists external IP addresses assigned to VMware workload VMs in a given private cloud.

pub async fn fetch_network_policy_external_addresses( &mut self, request: impl IntoRequest<FetchNetworkPolicyExternalAddressesRequest>, ) -> Result<Response<FetchNetworkPolicyExternalAddressesResponse>, Status>

Lists external IP addresses assigned to VMware workload VMs within the scope of the given network policy.

pub async fn get_external_address( &mut self, request: impl IntoRequest<GetExternalAddressRequest>, ) -> Result<Response<ExternalAddress>, Status>

Gets details of a single external IP address.

pub async fn create_external_address( &mut self, request: impl IntoRequest<CreateExternalAddressRequest>, ) -> Result<Response<Operation>, Status>

Creates a new ExternalAddress resource in a given private cloud. The network policy that corresponds to the private cloud must have the external IP address network service enabled (NetworkPolicy.external_ip).

pub async fn update_external_address( &mut self, request: impl IntoRequest<UpdateExternalAddressRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single external IP address. Only fields specified in update_mask are applied.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_external_address( &mut self, request: impl IntoRequest<DeleteExternalAddressRequest>, ) -> Result<Response<Operation>, Status>

Deletes a single external IP address. When you delete an external IP address, connectivity between the external IP address and the corresponding internal IP address is lost.

pub async fn list_subnets( &mut self, request: impl IntoRequest<ListSubnetsRequest>, ) -> Result<Response<ListSubnetsResponse>, Status>

Lists subnets in a given private cloud.

pub async fn get_subnet( &mut self, request: impl IntoRequest<GetSubnetRequest>, ) -> Result<Response<Subnet>, Status>

Gets details of a single subnet.

pub async fn update_subnet( &mut self, request: impl IntoRequest<UpdateSubnetRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single subnet. Only fields specified in update_mask are applied.

Note: This API is synchronous and always returns a successful google.longrunning.Operation (LRO). The returned LRO will only have done and response fields.

pub async fn list_external_access_rules( &mut self, request: impl IntoRequest<ListExternalAccessRulesRequest>, ) -> Result<Response<ListExternalAccessRulesResponse>, Status>

Lists ExternalAccessRule resources in the specified network policy.

pub async fn get_external_access_rule( &mut self, request: impl IntoRequest<GetExternalAccessRuleRequest>, ) -> Result<Response<ExternalAccessRule>, Status>

Gets details of a single external access rule.

pub async fn create_external_access_rule( &mut self, request: impl IntoRequest<CreateExternalAccessRuleRequest>, ) -> Result<Response<Operation>, Status>

Creates a new external access rule in a given network policy.

pub async fn update_external_access_rule( &mut self, request: impl IntoRequest<UpdateExternalAccessRuleRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single external access rule. Only fields specified in update_mask are applied.

pub async fn delete_external_access_rule( &mut self, request: impl IntoRequest<DeleteExternalAccessRuleRequest>, ) -> Result<Response<Operation>, Status>

Deletes a single external access rule.

pub async fn list_logging_servers( &mut self, request: impl IntoRequest<ListLoggingServersRequest>, ) -> Result<Response<ListLoggingServersResponse>, Status>

Lists logging servers configured for a given private cloud.

pub async fn get_logging_server( &mut self, request: impl IntoRequest<GetLoggingServerRequest>, ) -> Result<Response<LoggingServer>, Status>

Gets details of a logging server.

pub async fn create_logging_server( &mut self, request: impl IntoRequest<CreateLoggingServerRequest>, ) -> Result<Response<Operation>, Status>

Create a new logging server for a given private cloud.

pub async fn update_logging_server( &mut self, request: impl IntoRequest<UpdateLoggingServerRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of a single logging server. Only fields specified in update_mask are applied.

pub async fn delete_logging_server( &mut self, request: impl IntoRequest<DeleteLoggingServerRequest>, ) -> Result<Response<Operation>, Status>

Deletes a single logging server.

pub async fn list_node_types( &mut self, request: impl IntoRequest<ListNodeTypesRequest>, ) -> Result<Response<ListNodeTypesResponse>, Status>

Lists node types

pub async fn get_node_type( &mut self, request: impl IntoRequest<GetNodeTypeRequest>, ) -> Result<Response<NodeType>, Status>

Gets details of a single NodeType.

pub async fn show_nsx_credentials( &mut self, request: impl IntoRequest<ShowNsxCredentialsRequest>, ) -> Result<Response<Credentials>, Status>

Gets details of credentials for NSX appliance.

pub async fn show_vcenter_credentials( &mut self, request: impl IntoRequest<ShowVcenterCredentialsRequest>, ) -> Result<Response<Credentials>, Status>

Gets details of credentials for Vcenter appliance.

pub async fn reset_nsx_credentials( &mut self, request: impl IntoRequest<ResetNsxCredentialsRequest>, ) -> Result<Response<Operation>, Status>

Resets credentials of the NSX appliance.

pub async fn reset_vcenter_credentials( &mut self, request: impl IntoRequest<ResetVcenterCredentialsRequest>, ) -> Result<Response<Operation>, Status>

Resets credentials of the Vcenter appliance.

pub async fn get_dns_forwarding( &mut self, request: impl IntoRequest<GetDnsForwardingRequest>, ) -> Result<Response<DnsForwarding>, Status>

Gets details of the DnsForwarding config.

pub async fn update_dns_forwarding( &mut self, request: impl IntoRequest<UpdateDnsForwardingRequest>, ) -> Result<Response<Operation>, Status>

Updates the parameters of the DnsForwarding config, like associated domains. Only fields specified in update_mask are applied.

pub async fn get_network_peering( &mut self, request: impl IntoRequest<GetNetworkPeeringRequest>, ) -> Result<Response<NetworkPeering>, Status>

Retrieves a NetworkPeering resource by its resource name. The resource contains details of the network peering, such as peered networks, import and export custom route configurations, and peering state. NetworkPeering is a global resource and location can only be global.

pub async fn list_network_peerings( &mut self, request: impl IntoRequest<ListNetworkPeeringsRequest>, ) -> Result<Response<ListNetworkPeeringsResponse>, Status>

Lists NetworkPeering resources in a given project. NetworkPeering is a global resource and location can only be global.

pub async fn create_network_peering( &mut self, request: impl IntoRequest<CreateNetworkPeeringRequest>, ) -> Result<Response<Operation>, Status>

Creates a new network peering between the peer network and VMware Engine network provided in a NetworkPeering resource. NetworkPeering is a global resource and location can only be global.

pub async fn delete_network_peering( &mut self, request: impl IntoRequest<DeleteNetworkPeeringRequest>, ) -> Result<Response<Operation>, Status>

Deletes a NetworkPeering resource. When a network peering is deleted for a VMware Engine network, the peer network becomes inaccessible to that VMware Engine network. NetworkPeering is a global resource and location can only be global.

pub async fn update_network_peering( &mut self, request: impl IntoRequest<UpdateNetworkPeeringRequest>, ) -> Result<Response<Operation>, Status>

Modifies a NetworkPeering resource. Only the description field can be updated. Only fields specified in updateMask are applied. NetworkPeering is a global resource and location can only be global.

pub async fn list_peering_routes( &mut self, request: impl IntoRequest<ListPeeringRoutesRequest>, ) -> Result<Response<ListPeeringRoutesResponse>, Status>

Lists the network peering routes exchanged over a peering connection. NetworkPeering is a global resource and location can only be global.

pub async fn create_hcx_activation_key( &mut self, request: impl IntoRequest<CreateHcxActivationKeyRequest>, ) -> Result<Response<Operation>, Status>

Creates a new HCX activation key in a given private cloud.

pub async fn list_hcx_activation_keys( &mut self, request: impl IntoRequest<ListHcxActivationKeysRequest>, ) -> Result<Response<ListHcxActivationKeysResponse>, Status>

Lists HcxActivationKey resources in a given private cloud.

pub async fn get_hcx_activation_key( &mut self, request: impl IntoRequest<GetHcxActivationKeyRequest>, ) -> Result<Response<HcxActivationKey>, Status>

Retrieves a HcxActivationKey resource by its resource name.

pub async fn get_network_policy( &mut self, request: impl IntoRequest<GetNetworkPolicyRequest>, ) -> Result<Response<NetworkPolicy>, Status>

Retrieves a NetworkPolicy resource by its resource name.

pub async fn list_network_policies( &mut self, request: impl IntoRequest<ListNetworkPoliciesRequest>, ) -> Result<Response<ListNetworkPoliciesResponse>, Status>

Lists NetworkPolicy resources in a specified project and location.

pub async fn create_network_policy( &mut self, request: impl IntoRequest<CreateNetworkPolicyRequest>, ) -> Result<Response<Operation>, Status>

Creates a new network policy in a given VMware Engine network of a project and location (region). A new network policy cannot be created if another network policy already exists in the same scope.

pub async fn update_network_policy( &mut self, request: impl IntoRequest<UpdateNetworkPolicyRequest>, ) -> Result<Response<Operation>, Status>

Modifies a NetworkPolicy resource. Only the following fields can be updated: internet_access, external_ip, edge_services_cidr. Only fields specified in updateMask are applied. When updating a network policy, the external IP network service can only be disabled if there are no external IP addresses present in the scope of the policy. Also, a NetworkService cannot be updated when NetworkService.state is set to RECONCILING.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_network_policy( &mut self, request: impl IntoRequest<DeleteNetworkPolicyRequest>, ) -> Result<Response<Operation>, Status>

Deletes a NetworkPolicy resource. A network policy cannot be deleted when NetworkService.state is set to RECONCILING for either its external IP or internet access service.

pub async fn list_management_dns_zone_bindings( &mut self, request: impl IntoRequest<ListManagementDnsZoneBindingsRequest>, ) -> Result<Response<ListManagementDnsZoneBindingsResponse>, Status>

Lists Consumer VPCs bound to Management DNS Zone of a given private cloud.

pub async fn get_management_dns_zone_binding( &mut self, request: impl IntoRequest<GetManagementDnsZoneBindingRequest>, ) -> Result<Response<ManagementDnsZoneBinding>, Status>

Retrieves a ‘ManagementDnsZoneBinding’ resource by its resource name.

pub async fn create_management_dns_zone_binding( &mut self, request: impl IntoRequest<CreateManagementDnsZoneBindingRequest>, ) -> Result<Response<Operation>, Status>

Creates a new ManagementDnsZoneBinding resource in a private cloud. This RPC creates the DNS binding and the resource that represents the DNS binding of the consumer VPC network to the management DNS zone. A management DNS zone is the Cloud DNS cross-project binding zone that VMware Engine creates for each private cloud. It contains FQDNs and corresponding IP addresses for the private cloud’s ESXi hosts and management VM appliances like vCenter and NSX Manager.

pub async fn update_management_dns_zone_binding( &mut self, request: impl IntoRequest<UpdateManagementDnsZoneBindingRequest>, ) -> Result<Response<Operation>, Status>

Updates a ManagementDnsZoneBinding resource. Only fields specified in update_mask are applied.

pub async fn delete_management_dns_zone_binding( &mut self, request: impl IntoRequest<DeleteManagementDnsZoneBindingRequest>, ) -> Result<Response<Operation>, Status>

Deletes a ManagementDnsZoneBinding resource. When a management DNS zone binding is deleted, the corresponding consumer VPC network is no longer bound to the management DNS zone.

pub async fn repair_management_dns_zone_binding( &mut self, request: impl IntoRequest<RepairManagementDnsZoneBindingRequest>, ) -> Result<Response<Operation>, Status>

Retries to create a ManagementDnsZoneBinding resource that is in failed state.

pub async fn create_vmware_engine_network( &mut self, request: impl IntoRequest<CreateVmwareEngineNetworkRequest>, ) -> Result<Response<Operation>, Status>

Creates a new VMware Engine network that can be used by a private cloud.

pub async fn update_vmware_engine_network( &mut self, request: impl IntoRequest<UpdateVmwareEngineNetworkRequest>, ) -> Result<Response<Operation>, Status>

Modifies a VMware Engine network resource. Only the following fields can be updated: description. Only fields specified in updateMask are applied.

pub async fn delete_vmware_engine_network( &mut self, request: impl IntoRequest<DeleteVmwareEngineNetworkRequest>, ) -> Result<Response<Operation>, Status>

Deletes a VmwareEngineNetwork resource. You can only delete a VMware Engine network after all resources that refer to it are deleted. For example, a private cloud, a network peering, and a network policy can all refer to the same VMware Engine network.

pub async fn get_vmware_engine_network( &mut self, request: impl IntoRequest<GetVmwareEngineNetworkRequest>, ) -> Result<Response<VmwareEngineNetwork>, Status>

Retrieves a VmwareEngineNetwork resource by its resource name. The resource contains details of the VMware Engine network, such as its VMware Engine network type, peered networks in a service project, and state (for example, CREATING, ACTIVE, DELETING).

pub async fn list_vmware_engine_networks( &mut self, request: impl IntoRequest<ListVmwareEngineNetworksRequest>, ) -> Result<Response<ListVmwareEngineNetworksResponse>, Status>

Lists VmwareEngineNetwork resources in a given project and location.

pub async fn create_private_connection( &mut self, request: impl IntoRequest<CreatePrivateConnectionRequest>, ) -> Result<Response<Operation>, Status>

Creates a new private connection that can be used for accessing private Clouds.

pub async fn get_private_connection( &mut self, request: impl IntoRequest<GetPrivateConnectionRequest>, ) -> Result<Response<PrivateConnection>, Status>

Retrieves a PrivateConnection resource by its resource name. The resource contains details of the private connection, such as connected network, routing mode and state.

pub async fn list_private_connections( &mut self, request: impl IntoRequest<ListPrivateConnectionsRequest>, ) -> Result<Response<ListPrivateConnectionsResponse>, Status>

Lists PrivateConnection resources in a given project and location.

pub async fn update_private_connection( &mut self, request: impl IntoRequest<UpdatePrivateConnectionRequest>, ) -> Result<Response<Operation>, Status>

Modifies a PrivateConnection resource. Only description and routing_mode fields can be updated. Only fields specified in updateMask are applied.

pub async fn delete_private_connection( &mut self, request: impl IntoRequest<DeletePrivateConnectionRequest>, ) -> Result<Response<Operation>, Status>

Deletes a PrivateConnection resource. When a private connection is deleted for a VMware Engine network, the connected network becomes inaccessible to that VMware Engine network.

pub async fn list_private_connection_peering_routes( &mut self, request: impl IntoRequest<ListPrivateConnectionPeeringRoutesRequest>, ) -> Result<Response<ListPrivateConnectionPeeringRoutesResponse>, Status>

Lists the private connection routes exchanged over a peering connection.

pub async fn grant_dns_bind_permission( &mut self, request: impl IntoRequest<GrantDnsBindPermissionRequest>, ) -> Result<Response<Operation>, Status>

Grants the bind permission to the customer provided principal(user / service account) to bind their DNS zone with the intranet VPC associated with the project. DnsBindPermission is a global resource and location can only be global.

pub async fn get_dns_bind_permission( &mut self, request: impl IntoRequest<GetDnsBindPermissionRequest>, ) -> Result<Response<DnsBindPermission>, Status>

Gets all the principals having bind permission on the intranet VPC associated with the consumer project granted by the Grant API. DnsBindPermission is a global resource and location can only be global.

pub async fn revoke_dns_bind_permission( &mut self, request: impl IntoRequest<RevokeDnsBindPermissionRequest>, ) -> Result<Response<Operation>, Status>

Revokes the bind permission from the customer provided principal(user / service account) on the intranet VPC associated with the consumer project. DnsBindPermission is a global resource and location can only be global.