Enum google_api_proto::google::cloud::websecurityscanner::v1::scan_config_error::Code

#[repr(i32)]
pub enum Code {

    Unspecified = 0,
    InternalError = 1,
    AppengineApiBackendError = 2,
    AppengineApiNotAccessible = 3,
    AppengineDefaultHostMissing = 4,
    CannotUseGoogleComAccount = 6,
    CannotUseOwnerAccount = 7,
    ComputeApiBackendError = 8,
    ComputeApiNotAccessible = 9,
    CustomLoginUrlDoesNotBelongToCurrentProject = 10,
    CustomLoginUrlMalformed = 11,
    CustomLoginUrlMappedToNonRoutableAddress = 12,
    CustomLoginUrlMappedToUnreservedAddress = 13,
    CustomLoginUrlHasNonRoutableIpAddress = 14,
    CustomLoginUrlHasUnreservedIpAddress = 15,
    DuplicateScanName = 16,
    InvalidFieldValue = 18,
    FailedToAuthenticateToTarget = 19,
    FindingTypeUnspecified = 20,
    ForbiddenToScanCompute = 21,
    ForbiddenUpdateToManagedScan = 43,
    MalformedFilter = 22,
    MalformedResourceName = 23,
    ProjectInactive = 24,
    RequiredField = 25,
    ResourceNameInconsistent = 26,
    ScanAlreadyRunning = 27,
    ScanNotRunning = 28,
    SeedUrlDoesNotBelongToCurrentProject = 29,
    SeedUrlMalformed = 30,
    SeedUrlMappedToNonRoutableAddress = 31,
    SeedUrlMappedToUnreservedAddress = 32,
    SeedUrlHasNonRoutableIpAddress = 33,
    SeedUrlHasUnreservedIpAddress = 35,
    ServiceAccountNotConfigured = 36,
    TooManyScans = 37,
    UnableToResolveProjectInfo = 38,
    UnsupportedBlacklistPatternFormat = 39,
    UnsupportedFilter = 40,
    UnsupportedFindingType = 41,
    UnsupportedUrlScheme = 42,

}

Output only. Defines an error reason code. Next id: 44

Variants

Unspecified = 0

There is no error.

InternalError = 1

Indicates an internal server error. Please DO NOT USE THIS ERROR CODE unless the root cause is truly unknown.

AppengineApiBackendError = 2

One of the seed URLs is an App Engine URL but we cannot validate the scan settings due to an App Engine API backend error.

AppengineApiNotAccessible = 3

One of the seed URLs is an App Engine URL but we cannot access the App Engine API to validate scan settings.

AppengineDefaultHostMissing = 4

One of the seed URLs is an App Engine URL but the Default Host of the App Engine is not set.

CannotUseGoogleComAccount = 6

Google corporate accounts can not be used for scanning.

CannotUseOwnerAccount = 7

The account of the scan creator can not be used for scanning.

ComputeApiBackendError = 8

This scan targets Compute Engine, but we cannot validate scan settings due to a Compute Engine API backend error.

ComputeApiNotAccessible = 9

This scan targets Compute Engine, but we cannot access the Compute Engine API to validate the scan settings.

CustomLoginUrlDoesNotBelongToCurrentProject = 10

The Custom Login URL does not belong to the current project.

CustomLoginUrlMalformed = 11

The Custom Login URL is malformed (can not be parsed).

CustomLoginUrlMappedToNonRoutableAddress = 12

The Custom Login URL is mapped to a non-routable IP address in DNS.

CustomLoginUrlMappedToUnreservedAddress = 13

The Custom Login URL is mapped to an IP address which is not reserved for the current project.

CustomLoginUrlHasNonRoutableIpAddress = 14

The Custom Login URL has a non-routable IP address.

CustomLoginUrlHasUnreservedIpAddress = 15

The Custom Login URL has an IP address which is not reserved for the current project.

DuplicateScanName = 16

Another scan with the same name (case-sensitive) already exists.

InvalidFieldValue = 18

A field is set to an invalid value.

FailedToAuthenticateToTarget = 19

There was an error trying to authenticate to the scan target.

FindingTypeUnspecified = 20

Finding type value is not specified in the list findings request.

ForbiddenToScanCompute = 21

Scan targets Compute Engine, yet current project was not whitelisted for Google Compute Engine Scanning Alpha access.

ForbiddenUpdateToManagedScan = 43

User tries to update managed scan

MalformedFilter = 22

The supplied filter is malformed. For example, it can not be parsed, does not have a filter type in expression, or the same filter type appears more than once.

MalformedResourceName = 23

The supplied resource name is malformed (can not be parsed).

ProjectInactive = 24

The current project is not in an active state.

RequiredField = 25

A required field is not set.

ResourceNameInconsistent = 26

Project id, scanconfig id, scanrun id, or finding id are not consistent with each other in resource name.

ScanAlreadyRunning = 27

The scan being requested to start is already running.

ScanNotRunning = 28

The scan that was requested to be stopped is not running.

SeedUrlDoesNotBelongToCurrentProject = 29

One of the seed URLs does not belong to the current project.

SeedUrlMalformed = 30

One of the seed URLs is malformed (can not be parsed).

SeedUrlMappedToNonRoutableAddress = 31

One of the seed URLs is mapped to a non-routable IP address in DNS.

SeedUrlMappedToUnreservedAddress = 32

One of the seed URLs is mapped to an IP address which is not reserved for the current project.

SeedUrlHasNonRoutableIpAddress = 33

One of the seed URLs has on-routable IP address.

SeedUrlHasUnreservedIpAddress = 35

One of the seed URLs has an IP address that is not reserved for the current project.

ServiceAccountNotConfigured = 36

The Web Security Scanner service account is not configured under the project.

TooManyScans = 37

A project has reached the maximum number of scans.

UnableToResolveProjectInfo = 38

Resolving the details of the current project fails.

UnsupportedBlacklistPatternFormat = 39

One or more blacklist patterns were in the wrong format.

UnsupportedFilter = 40

The supplied filter is not supported.

UnsupportedFindingType = 41

The supplied finding type is not supported. For example, we do not provide findings of the given finding type.

UnsupportedUrlScheme = 42

The URL scheme of one or more of the supplied URLs is not supported.

Implementations

impl Code

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Code.

pub fn from_i32(value: i32) -> Option<Code>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Code, or None if value is not a valid variant.

impl Code

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.