Module google_api_proto::google::cloud::websecurityscanner::v1beta

Modules

scan_config

Nested message and enum types in ScanConfig.

scan_config_error

Nested message and enum types in ScanConfigError.

scan_run

Nested message and enum types in ScanRun.

scan_run_error_trace

Nested message and enum types in ScanRunErrorTrace.

scan_run_warning_trace

Nested message and enum types in ScanRunWarningTrace.

vulnerable_headers

Nested message and enum types in VulnerableHeaders.

web_security_scanner_client

Generated client implementations.

Structs

CrawledUrl

A CrawledUrl resource represents a URL that was crawled during a ScanRun. Web Security Scanner Service crawls the web applications, following all links within the scope of sites, to find the URLs to test against.

CreateScanConfigRequest

Request for the CreateScanConfig method.

DeleteScanConfigRequest

Request for the DeleteScanConfig method.

Finding

A Finding resource represents a vulnerability instance identified during a ScanRun.

FindingTypeStats

A FindingTypeStats resource represents stats regarding a specific FindingType of Findings under a given ScanRun.

Form

! Information about a vulnerability with an HTML.

GetFindingRequest

Request for the GetFinding method.

GetScanConfigRequest

Request for the GetScanConfig method.

GetScanRunRequest

Request for the GetScanRun method.

ListCrawledUrlsRequest

Request for the ListCrawledUrls method.

ListCrawledUrlsResponse

Response for the ListCrawledUrls method.

ListFindingTypeStatsRequest

Request for the ListFindingTypeStats method.

ListFindingTypeStatsResponse

Response for the ListFindingTypeStats method.

ListFindingsRequest

Request for the ListFindings method.

ListFindingsResponse

Response for the ListFindings method.

ListScanConfigsRequest

Request for the ListScanConfigs method.

ListScanConfigsResponse

Response for the ListScanConfigs method.

ListScanRunsRequest

Request for the ListScanRuns method.

ListScanRunsResponse

Response for the ListScanRuns method.

OutdatedLibrary

Information reported for an outdated library.

ScanConfig

A ScanConfig resource contains the configurations to launch a scan.

ScanConfigError

Defines a custom error message used by CreateScanConfig and UpdateScanConfig APIs when scan configuration validation fails. It is also reported as part of a ScanRunErrorTrace message if scan validation fails due to a scan configuration error.

ScanRun

A ScanRun is a output-only resource representing an actual run of the scan. Next id: 12

ScanRunErrorTrace

Output only. Defines an error trace message for a ScanRun.

ScanRunWarningTrace

Output only. Defines a warning trace message for ScanRun. Warning traces provide customers with useful information that helps make the scanning process more effective.

StartScanRunRequest

Request for the StartScanRun method.

StopScanRunRequest

Request for the StopScanRun method.

UpdateScanConfigRequest

Request for the UpdateScanConfigRequest method.

ViolatingResource

Information regarding any resource causing the vulnerability such as JavaScript sources, image, audio files, etc.

VulnerableHeaders

Information about vulnerable or missing HTTP Headers.

VulnerableParameters

Information about vulnerable request parameters.

Xss

Information reported for an XSS.