Struct google_api_proto::google::cloud::workstations::v1beta::WorkstationConfig

pub struct WorkstationConfig {

    pub name: String,
    pub display_name: String,
    pub uid: String,
    pub reconciling: bool,
    pub annotations: BTreeMap<String, String>,
    pub labels: BTreeMap<String, String>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub delete_time: Option<Timestamp>,
    pub etag: String,
    pub idle_timeout: Option<Duration>,
    pub running_timeout: Option<Duration>,
    pub host: Option<Host>,
    pub persistent_directories: Vec<PersistentDirectory>,
    pub ephemeral_directories: Vec<EphemeralDirectory>,
    pub container: Option<Container>,
    pub encryption_key: Option<CustomerEncryptionKey>,
    pub readiness_checks: Vec<ReadinessCheck>,
    pub replica_zones: Vec<String>,
    pub degraded: bool,
    pub conditions: Vec<Status>,
    pub enable_audit_agent: bool,

}

A workstation configuration resource in the Cloud Workstations API.

Workstation configurations act as templates for workstations. The workstation configuration defines details such as the workstation virtual machine (VM) instance type, persistent storage, container image defining environment, which IDE or Code Editor to use, and more. Administrators and platform teams can also use Identity and Access Management (IAM) rules to grant access to teams or to individual developers.

Fields
name: String

Full name of this workstation configuration.

display_name: String

Optional. Human-readable name for this workstation configuration.

uid: String

Output only. A system-assigned unique identifier for this workstation configuration.

reconciling: bool

Output only. Indicates whether this workstation configuration is currently being updated to match its intended state.

annotations: BTreeMap<String, String>

Optional. Client-specified annotations.

labels: BTreeMap<String, String>

Optional. Labels that are applied to the workstation configuration and that are also propagated to the underlying Compute Engine resources.

create_time: Option<Timestamp>

Output only. Time when this workstation configuration was created.

update_time: Option<Timestamp>

Output only. Time when this workstation configuration was most recently updated.

delete_time: Option<Timestamp>

Output only. Time when this workstation configuration was soft-deleted.

etag: String

Optional. Checksum computed by the server. May be sent on update and delete requests to make sure that the client has an up-to-date value before proceeding.

idle_timeout: Option<Duration>

Optional. Number of seconds to wait before automatically stopping a workstation after it last received user traffic.

A value of "0s" indicates that Cloud Workstations VMs created with this configuration should never time out due to idleness. Provide duration terminated by s for seconds—for example, "7200s" (2 hours). The default is "1200s" (20 minutes).

running_timeout: Option<Duration>

Optional. Number of seconds that a workstation can run until it is automatically shut down. We recommend that workstations be shut down daily to reduce costs and so that security updates can be applied upon restart. The [idle_timeout][google.cloud.workstations.v1beta.WorkstationConfig.idle_timeout] and [running_timeout][google.cloud.workstations.v1beta.WorkstationConfig.running_timeout] fields are independent of each other. Note that the [running_timeout][google.cloud.workstations.v1beta.WorkstationConfig.running_timeout] field shuts down VMs after the specified time, regardless of whether or not the VMs are idle.

Provide duration terminated by s for seconds—for example, "54000s" (15 hours). Defaults to "43200s" (12 hours). A value of "0s" indicates that workstations using this configuration should never time out. If [encryption_key][google.cloud.workstations.v1beta.WorkstationConfig.encryption_key] is set, it must be greater than "0s" and less than "86400s" (24 hours).

Warning: A value of "0s" indicates that Cloud Workstations VMs created with this configuration have no maximum running time. This is strongly discouraged because you incur costs and will not pick up security updates.

host: Option<Host>

Optional. Runtime host for the workstation.

persistent_directories: Vec<PersistentDirectory>

Optional. Directories to persist across workstation sessions.

ephemeral_directories: Vec<EphemeralDirectory>

Optional. Ephemeral directories which won’t persist across workstation sessions.

container: Option<Container>

Optional. Container that runs upon startup for each workstation using this workstation configuration.

encryption_key: Option<CustomerEncryptionKey>

Immutable. Encrypts resources of this workstation configuration using a customer-managed encryption key (CMEK).

If specified, the boot disk of the Compute Engine instance and the persistent disk are encrypted using this encryption key. If this field is not set, the disks are encrypted using a generated key. Customer-managed encryption keys do not protect disk metadata.

If the customer-managed encryption key is rotated, when the workstation instance is stopped, the system attempts to recreate the persistent disk with the new version of the key. Be sure to keep older versions of the key until the persistent disk is recreated. Otherwise, data on the persistent disk might be lost.

If the encryption key is revoked, the workstation session automatically stops within 7 hours.

Immutable after the workstation configuration is created.

readiness_checks: Vec<ReadinessCheck>

Optional. Readiness checks to perform when starting a workstation using this workstation configuration. Mark a workstation as running only after all specified readiness checks return 200 status codes.

replica_zones: Vec<String>

Optional. Immutable. Specifies the zones used to replicate the VM and disk resources within the region. If set, exactly two zones within the workstation cluster’s region must be specified—for example, \['us-central1-a', 'us-central1-f'\]. If this field is empty, two default zones within the region are used.

Immutable after the workstation configuration is created.

degraded: bool

Output only. Whether this resource is degraded, in which case it may require user action to restore full functionality. See also the [conditions][google.cloud.workstations.v1beta.WorkstationConfig.conditions] field.

conditions: Vec<Status>

Output only. Status conditions describing the current resource state.

enable_audit_agent: bool

Optional. Whether to enable Linux auditd logging on the workstation. When enabled, a service account must also be specified that has logging.buckets.write permission on the project. Operating system audit logging is distinct from Cloud Audit Logs.