Module google_api_proto::google::cloud::workstations::v1beta::workstation_config

Nested message and enum types in WorkstationConfig.

Modules

ephemeral_directory

Nested message and enum types in EphemeralDirectory.

host

Nested message and enum types in Host.

persistent_directory

Nested message and enum types in PersistentDirectory.

Structs

Container

A Docker container.

CustomerEncryptionKey

A customer-managed encryption key (CMEK) for the Compute Engine resources of the associated workstation configuration. Specify the name of your Cloud KMS encryption key and the default service account. We recommend that you use a separate service account and follow Cloud KMS best practices.

EphemeralDirectory

An ephemeral directory which won’t persist across workstation sessions. It is freshly created on every workstation start operation.

Host

Runtime host for a workstation.

PersistentDirectory

A directory to persist across workstation sessions.

ReadinessCheck

A readiness check to be performed on a workstation.