Struct google_api_proto::google::cloud::workstations::v1beta::workstations_client::WorkstationsClient

pub struct WorkstationsClient<T> { /* private fields */ }

Service for interacting with Cloud Workstations.

Implementations

impl<T> WorkstationsClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> WorkstationsClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_workstation_cluster( &mut self, request: impl IntoRequest<GetWorkstationClusterRequest>, ) -> Result<Response<WorkstationCluster>, Status>

Returns the requested workstation cluster.

pub async fn list_workstation_clusters( &mut self, request: impl IntoRequest<ListWorkstationClustersRequest>, ) -> Result<Response<ListWorkstationClustersResponse>, Status>

Returns all workstation clusters in the specified location.

pub async fn create_workstation_cluster( &mut self, request: impl IntoRequest<CreateWorkstationClusterRequest>, ) -> Result<Response<Operation>, Status>

Creates a new workstation cluster.

pub async fn update_workstation_cluster( &mut self, request: impl IntoRequest<UpdateWorkstationClusterRequest>, ) -> Result<Response<Operation>, Status>

Updates an existing workstation cluster.

pub async fn delete_workstation_cluster( &mut self, request: impl IntoRequest<DeleteWorkstationClusterRequest>, ) -> Result<Response<Operation>, Status>

Deletes the specified workstation cluster.

pub async fn get_workstation_config( &mut self, request: impl IntoRequest<GetWorkstationConfigRequest>, ) -> Result<Response<WorkstationConfig>, Status>

Returns the requested workstation configuration.

pub async fn list_workstation_configs( &mut self, request: impl IntoRequest<ListWorkstationConfigsRequest>, ) -> Result<Response<ListWorkstationConfigsResponse>, Status>

Returns all workstation configurations in the specified cluster.

pub async fn list_usable_workstation_configs( &mut self, request: impl IntoRequest<ListUsableWorkstationConfigsRequest>, ) -> Result<Response<ListUsableWorkstationConfigsResponse>, Status>

Returns all workstation configurations in the specified cluster on which the caller has the “workstations.workstation.create” permission.

pub async fn create_workstation_config( &mut self, request: impl IntoRequest<CreateWorkstationConfigRequest>, ) -> Result<Response<Operation>, Status>

Creates a new workstation configuration.

pub async fn update_workstation_config( &mut self, request: impl IntoRequest<UpdateWorkstationConfigRequest>, ) -> Result<Response<Operation>, Status>

Updates an existing workstation configuration.

pub async fn delete_workstation_config( &mut self, request: impl IntoRequest<DeleteWorkstationConfigRequest>, ) -> Result<Response<Operation>, Status>

Deletes the specified workstation configuration.

pub async fn get_workstation( &mut self, request: impl IntoRequest<GetWorkstationRequest>, ) -> Result<Response<Workstation>, Status>

Returns the requested workstation.

pub async fn list_workstations( &mut self, request: impl IntoRequest<ListWorkstationsRequest>, ) -> Result<Response<ListWorkstationsResponse>, Status>

Returns all Workstations using the specified workstation configuration.

pub async fn list_usable_workstations( &mut self, request: impl IntoRequest<ListUsableWorkstationsRequest>, ) -> Result<Response<ListUsableWorkstationsResponse>, Status>

Returns all workstations using the specified workstation configuration on which the caller has the “workstations.workstations.use” permission.

pub async fn create_workstation( &mut self, request: impl IntoRequest<CreateWorkstationRequest>, ) -> Result<Response<Operation>, Status>

Creates a new workstation.

pub async fn update_workstation( &mut self, request: impl IntoRequest<UpdateWorkstationRequest>, ) -> Result<Response<Operation>, Status>

Updates an existing workstation.

pub async fn delete_workstation( &mut self, request: impl IntoRequest<DeleteWorkstationRequest>, ) -> Result<Response<Operation>, Status>

Deletes the specified workstation.

pub async fn start_workstation( &mut self, request: impl IntoRequest<StartWorkstationRequest>, ) -> Result<Response<Operation>, Status>

Starts running a workstation so that users can connect to it.

pub async fn stop_workstation( &mut self, request: impl IntoRequest<StopWorkstationRequest>, ) -> Result<Response<Operation>, Status>

Stops running a workstation, reducing costs.

pub async fn generate_access_token( &mut self, request: impl IntoRequest<GenerateAccessTokenRequest>, ) -> Result<Response<GenerateAccessTokenResponse>, Status>

Returns a short-lived credential that can be used to send authenticated and authorized traffic to a workstation.