Module google_api_proto::google::container::v1

Modules

advanced_datapath_observability_config

Nested message and enum types in AdvancedDatapathObservabilityConfig.

autopilot_compatibility_issue

Nested message and enum types in AutopilotCompatibilityIssue.

binary_authorization

Nested message and enum types in BinaryAuthorization.

blue_green_settings

Nested message and enum types in BlueGreenSettings.

cloud_run_config

Nested message and enum types in CloudRunConfig.

cluster

Nested message and enum types in Cluster.

cluster_autoscaling

Nested message and enum types in ClusterAutoscaling.

cluster_manager_client

Generated client implementations.

containerd_config

Nested message and enum types in ContainerdConfig.

database_encryption

Nested message and enum types in DatabaseEncryption.

dns_config

Nested message and enum types in DNSConfig.

enterprise_config

Nested message and enum types in EnterpriseConfig.

gateway_api_config

Nested message and enum types in GatewayAPIConfig.

gpu_driver_installation_config

Nested message and enum types in GPUDriverInstallationConfig.

gpu_sharing_config

Nested message and enum types in GPUSharingConfig.

linux_node_config

Nested message and enum types in LinuxNodeConfig.

logging_component_config

Nested message and enum types in LoggingComponentConfig.

logging_variant_config

Nested message and enum types in LoggingVariantConfig.

maintenance_exclusion_options

Nested message and enum types in MaintenanceExclusionOptions.

maintenance_window

Nested message and enum types in MaintenanceWindow.

master_authorized_networks_config

Nested message and enum types in MasterAuthorizedNetworksConfig.

monitoring_component_config

Nested message and enum types in MonitoringComponentConfig.

network_config

Nested message and enum types in NetworkConfig.

network_policy

Nested message and enum types in NetworkPolicy.

node_network_config

Nested message and enum types in NodeNetworkConfig.

node_pool

Nested message and enum types in NodePool.

node_pool_autoscaling

Nested message and enum types in NodePoolAutoscaling.

node_taint

Nested message and enum types in NodeTaint.

notification_config

Nested message and enum types in NotificationConfig.

operation

Nested message and enum types in Operation.

operation_progress

Nested message and enum types in OperationProgress.

release_channel

Nested message and enum types in ReleaseChannel.

reservation_affinity

Nested message and enum types in ReservationAffinity.

resource_usage_export_config

Nested message and enum types in ResourceUsageExportConfig.

sandbox_config

Nested message and enum types in SandboxConfig.

secondary_boot_disk

Nested message and enum types in SecondaryBootDisk.

security_posture_config

Nested message and enum types in SecurityPostureConfig.

server_config

Nested message and enum types in ServerConfig.

set_master_auth_request

Nested message and enum types in SetMasterAuthRequest.

sole_tenant_config

Nested message and enum types in SoleTenantConfig.

status_condition

Nested message and enum types in StatusCondition.

time_window

Nested message and enum types in TimeWindow.

usable_subnetwork_secondary_range

Nested message and enum types in UsableSubnetworkSecondaryRange.

windows_node_config

Nested message and enum types in WindowsNodeConfig.

workload_metadata_config

Nested message and enum types in WorkloadMetadataConfig.

Structs

AcceleratorConfig

AcceleratorConfig represents a Hardware Accelerator request.

AdditionalNodeNetworkConfig

AdditionalNodeNetworkConfig is the configuration for additional node networks within the NodeNetworkConfig message

AdditionalPodNetworkConfig

AdditionalPodNetworkConfig is the configuration for additional pod networks within the NodeNetworkConfig message

AdditionalPodRangesConfig

AdditionalPodRangesConfig is the configuration for additional pod secondary ranges supporting the ClusterUpdate message.

AddonsConfig

Configuration for the addons that can be automatically spun up in the cluster, enabling additional functionality.

AdvancedDatapathObservabilityConfig

AdvancedDatapathObservabilityConfig specifies configuration of observability features of advanced datapath.

AdvancedMachineFeatures

Specifies options for controlling advanced machine features.

AuthenticatorGroupsConfig

Configuration for returning group information from authenticators.

AutoUpgradeOptions

AutoUpgradeOptions defines the set of options for the user to control how the Auto Upgrades will proceed.

Autopilot

Autopilot is the configuration for Autopilot settings on the cluster.

AutopilotCompatibilityIssue

AutopilotCompatibilityIssue contains information about a specific compatibility issue with Autopilot mode.

AutoprovisioningNodePoolDefaults

AutoprovisioningNodePoolDefaults contains defaults for a node pool created by NAP.

BestEffortProvisioning

Best effort provisioning.

BinaryAuthorization

Configuration for Binary Authorization.

BlueGreenSettings

Settings for blue-green upgrade.

CancelOperationRequest

CancelOperationRequest cancels a single operation.

CheckAutopilotCompatibilityRequest

CheckAutopilotCompatibilityRequest requests getting the blockers for the given operation in the cluster.

CheckAutopilotCompatibilityResponse

CheckAutopilotCompatibilityResponse has a list of compatibility issues.

ClientCertificateConfig

Configuration for client certificates on the cluster.

CloudRunConfig

Configuration options for the Cloud Run feature.

Cluster

A Google Kubernetes Engine cluster.

ClusterAutoscaling

ClusterAutoscaling contains global, per-cluster information required by Cluster Autoscaler to automatically adjust the size of the cluster and create/delete node pools based on the current needs.

ClusterUpdate

ClusterUpdate describes an update to the cluster. Exactly one update can be applied to a cluster with each request, so at most one field can be provided.

CompleteIpRotationRequest

CompleteIPRotationRequest moves the cluster master back into single-IP mode.

CompleteNodePoolUpgradeRequest

CompleteNodePoolUpgradeRequest sets the name of target node pool to complete upgrade.

ConfidentialNodes

ConfidentialNodes is configuration for the confidential nodes feature, which makes nodes run on confidential VMs.

ConfigConnectorConfig

Configuration options for the Config Connector add-on.

ContainerdConfig

ContainerdConfig contains configuration to customize containerd.

CostManagementConfig

Configuration for fine-grained cost management feature.

CreateClusterRequest

CreateClusterRequest creates a cluster.

CreateNodePoolRequest

CreateNodePoolRequest creates a node pool for a cluster.

DailyMaintenanceWindow

Time window specified for daily maintenance operations.

DatabaseEncryption

Configuration of etcd encryption.

DefaultSnatStatus

DefaultSnatStatus contains the desired state of whether default sNAT should be disabled on the cluster.

DeleteClusterRequest

DeleteClusterRequest deletes a cluster.

DeleteNodePoolRequest

DeleteNodePoolRequest deletes a node pool for a cluster.

DnsCacheConfig

Configuration for NodeLocal DNSCache

DnsConfig

DNSConfig contains the desired set of options for configuring clusterDNS.

EnterpriseConfig

EnterpriseConfig is the cluster enterprise configuration.

EphemeralStorageLocalSsdConfig

EphemeralStorageLocalSsdConfig contains configuration for the node ephemeral storage using Local SSDs.

FastSocket

Configuration of Fast Socket feature.

Fleet

Fleet is the fleet configuration for the cluster.

GatewayApiConfig

GatewayAPIConfig contains the desired config of Gateway API on this cluster.

GcePersistentDiskCsiDriverConfig

Configuration for the Compute Engine PD CSI driver.

GcfsConfig

GcfsConfig contains configurations of Google Container File System (image streaming).

GcpFilestoreCsiDriverConfig

Configuration for the GCP Filestore CSI driver.

GcsFuseCsiDriverConfig

Configuration for the Cloud Storage Fuse CSI driver.

GetClusterRequest

GetClusterRequest gets the settings of a cluster.

GetJsonWebKeysRequest

GetJSONWebKeysRequest gets the public component of the keys used by the cluster to sign token requests. This will be the jwks_uri for the discover document returned by getOpenIDConfig. See the OpenID Connect Discovery 1.0 specification for details.

GetJsonWebKeysResponse

GetJSONWebKeysResponse is a valid JSON Web Key Set as specififed in rfc 7517

GetNodePoolRequest

GetNodePoolRequest retrieves a node pool for a cluster.

GetOpenIdConfigRequest

GetOpenIDConfigRequest gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

GetOpenIdConfigResponse

GetOpenIDConfigResponse is an OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

GetOperationRequest

GetOperationRequest gets a single operation.

GetServerConfigRequest

Gets the current Kubernetes Engine service configuration.

GkeBackupAgentConfig

Configuration for the Backup for GKE Agent.

GpuDriverInstallationConfig

GPUDriverInstallationConfig specifies the version of GPU driver to be auto installed.

GpuSharingConfig

GPUSharingConfig represents the GPU sharing configuration for Hardware Accelerators.

HorizontalPodAutoscaling

Configuration options for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods.

HttpLoadBalancing

Configuration options for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster.

IdentityServiceConfig

IdentityServiceConfig is configuration for Identity Service which allows customers to use external identity providers with the K8S API

IlbSubsettingConfig

ILBSubsettingConfig contains the desired config of L4 Internal LoadBalancer subsetting on this cluster.

IntraNodeVisibilityConfig

IntraNodeVisibilityConfig contains the desired config of the intra-node visibility on this cluster.

IpAllocationPolicy

Configuration for controlling how IPs are allocated in the cluster.

Jwk

Jwk is a JSON Web Key as specified in RFC 7517

K8sBetaApiConfig

K8sBetaAPIConfig , configuration for beta APIs

KubernetesDashboard

Configuration for the Kubernetes Dashboard.

LegacyAbac

Configuration for the legacy Attribute Based Access Control authorization mode.

LinuxNodeConfig

Parameters that can be configured on Linux nodes.

ListClustersRequest

ListClustersRequest lists clusters.

ListClustersResponse

ListClustersResponse is the result of ListClustersRequest.

ListNodePoolsRequest

ListNodePoolsRequest lists the node pool(s) for a cluster.

ListNodePoolsResponse

ListNodePoolsResponse is the result of ListNodePoolsRequest.

ListOperationsRequest

ListOperationsRequest lists operations.

ListOperationsResponse

ListOperationsResponse is the result of ListOperationsRequest.

ListUsableSubnetworksRequest

ListUsableSubnetworksRequest requests the list of usable subnetworks available to a user for creating clusters.

ListUsableSubnetworksResponse

ListUsableSubnetworksResponse is the response of ListUsableSubnetworksRequest.

LocalNvmeSsdBlockConfig

LocalNvmeSsdBlockConfig contains configuration for using raw-block local NVMe SSDs

LoggingComponentConfig

LoggingComponentConfig is cluster logging component configuration.

LoggingConfig

LoggingConfig is cluster logging configuration.

LoggingVariantConfig

LoggingVariantConfig specifies the behaviour of the logging component.

MaintenanceExclusionOptions

Represents the Maintenance exclusion option.

MaintenancePolicy

MaintenancePolicy defines the maintenance policy to be used for the cluster.

MaintenanceWindow

MaintenanceWindow defines the maintenance window to be used for the cluster.

ManagedPrometheusConfig

ManagedPrometheusConfig defines the configuration for Google Cloud Managed Service for Prometheus.

MasterAuth

The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.

MasterAuthorizedNetworksConfig

Configuration options for the master authorized networks feature. Enabled master authorized networks will disallow all external traffic to access Kubernetes master through HTTPS except traffic from the given CIDR blocks, Google Compute Engine Public IPs and Google Prod IPs.

MaxPodsConstraint

Constraints applied to pods.

MeshCertificates

Configuration for issuance of mTLS keys and certificates to Kubernetes pods.

MonitoringComponentConfig

MonitoringComponentConfig is cluster monitoring component configuration.

MonitoringConfig

MonitoringConfig is cluster monitoring configuration.

NetworkConfig

NetworkConfig reports the relative names of network & subnetwork.

NetworkPolicy

Configuration options for the NetworkPolicy feature. https://kubernetes.io/docs/concepts/services-networking/networkpolicies/

NetworkPolicyConfig

Configuration for NetworkPolicy. This only tracks whether the addon is enabled or not on the Master, it does not track whether network policy is enabled for the nodes.

NetworkTags

Collection of Compute Engine network tags that can be applied to a node’s underlying VM instance.

NodeConfig

Parameters that describe the nodes in a cluster.

NodeConfigDefaults

Subset of NodeConfig message that has defaults.

NodeKubeletConfig

Node kubelet configs.

NodeLabels

Collection of node-level Kubernetes labels.

NodeManagement

NodeManagement defines the set of node management services turned on for the node pool.

NodeNetworkConfig

Parameters for node pool-level network config.

NodePool

NodePool contains the name and configuration for a cluster’s node pool. Node pools are a set of nodes (i.e. VM’s), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.

NodePoolAutoConfig

Node pool configs that apply to all auto-provisioned node pools in autopilot clusters and node auto-provisioning enabled clusters.

NodePoolAutoscaling

NodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

NodePoolDefaults

Subset of Nodepool message that has defaults.

NodePoolLoggingConfig

NodePoolLoggingConfig specifies logging configuration for nodepools.

NodeTaint

Kubernetes taint is composed of three fields: key, value, and effect. Effect can only be one of three types: NoSchedule, PreferNoSchedule or NoExecute.

NodeTaints

Collection of Kubernetes node taints.

NotificationConfig

NotificationConfig is the configuration of notifications.

Operation

This operation resource represents operations that may have happened or are happening on the cluster. All fields are output only.

OperationProgress

Information about operation (or operation stage) progress.

PodCidrOverprovisionConfig

[PRIVATE FIELD] Config for pod CIDR size overprovisioning.

PrivateClusterConfig

Configuration options for private clusters.

PrivateClusterMasterGlobalAccessConfig

Configuration for controlling master global access settings.

RangeInfo

RangeInfo contains the range name and the range utilization by this cluster.

RayClusterLoggingConfig

RayClusterLoggingConfig specifies configuration of Ray logging.

RayClusterMonitoringConfig

RayClusterMonitoringConfig specifies monitoring configuration for Ray clusters.

RayOperatorConfig

Configuration options for the Ray Operator add-on.

RecurringTimeWindow

Represents an arbitrary window of time that recurs.

ReleaseChannel

ReleaseChannel indicates which release channel a cluster is subscribed to. Release channels are arranged in order of risk.

ReservationAffinity

ReservationAffinity is the configuration of desired reservation which instances could take capacity from.

ResourceLabels

Collection of GCP labels.

ResourceLimit

Contains information about amount of some resource in the cluster. For memory, value should be in GB.

ResourceManagerTags

A map of resource manager tag keys and values to be attached to the nodes for managing Compute Engine firewalls using Network Firewall Policies. Tags must be according to specifications in https://cloud.google.com/vpc/docs/tags-firewalls-overview#specifications. A maximum of 5 tag key-value pairs can be specified. Existing tags will be replaced with new values.

ResourceUsageExportConfig

Configuration for exporting cluster resource usages.

RollbackNodePoolUpgradeRequest

RollbackNodePoolUpgradeRequest rollbacks the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.

SandboxConfig

SandboxConfig contains configurations of the sandbox to use for the node.

SecondaryBootDisk

SecondaryBootDisk represents a persistent disk attached to a node with special configurations based on its mode.

SecondaryBootDiskUpdateStrategy

SecondaryBootDiskUpdateStrategy is a placeholder which will be extended in the future to define different options for updating secondary boot disks.

SecurityBulletinEvent

SecurityBulletinEvent is a notification sent to customers when a security bulletin has been posted that they are vulnerable to.

SecurityPostureConfig

SecurityPostureConfig defines the flags needed to enable/disable features for the Security Posture API.

ServerConfig

Kubernetes Engine service configuration.

ServiceExternalIPsConfig

Config to block services with externalIPs field.

SetAddonsConfigRequest

SetAddonsConfigRequest sets the addons associated with the cluster.

SetLabelsRequest

SetLabelsRequest sets the Google Cloud Platform labels on a Google Container Engine cluster, which will in turn set them for Google Compute Engine resources used by that cluster

SetLegacyAbacRequest

SetLegacyAbacRequest enables or disables the ABAC authorization mechanism for a cluster.

SetLocationsRequest

SetLocationsRequest sets the locations of the cluster.

SetLoggingServiceRequest

SetLoggingServiceRequest sets the logging service of a cluster.

SetMaintenancePolicyRequest

SetMaintenancePolicyRequest sets the maintenance policy for a cluster.

SetMasterAuthRequest

SetMasterAuthRequest updates the admin password of a cluster.

SetMonitoringServiceRequest

SetMonitoringServiceRequest sets the monitoring service of a cluster.

SetNetworkPolicyRequest

SetNetworkPolicyRequest enables/disables network policy for a cluster.

SetNodePoolAutoscalingRequest

SetNodePoolAutoscalingRequest sets the autoscaler settings of a node pool.

SetNodePoolManagementRequest

SetNodePoolManagementRequest sets the node management properties of a node pool.

SetNodePoolSizeRequest

SetNodePoolSizeRequest sets the size of a node pool.

ShieldedInstanceConfig

A set of Shielded Instance options.

ShieldedNodes

Configuration of Shielded Nodes feature.

SoleTenantConfig

SoleTenantConfig contains the NodeAffinities to specify what shared sole tenant node groups should back the node pool.

StartIpRotationRequest

StartIPRotationRequest creates a new IP for the cluster and then performs a node upgrade on each node pool to point to the new IP.

StatefulHaConfig

Configuration for the Stateful HA add-on.

StatusCondition

StatusCondition describes why a cluster or a node pool has a certain status (e.g., ERROR or DEGRADED).

TimeWindow

Represents an arbitrary window of time.

UpdateClusterRequest

UpdateClusterRequest updates the settings of a cluster.

UpdateMasterRequest

UpdateMasterRequest updates the master of the cluster.

UpdateNodePoolRequest

UpdateNodePoolRequests update a node pool’s image and/or version.

UpgradeAvailableEvent

UpgradeAvailableEvent is a notification sent to customers when a new available version is released.

UpgradeEvent

UpgradeEvent is a notification sent to customers by the cluster server when a resource is upgrading.

UsableSubnetwork

UsableSubnetwork resource returns the subnetwork name, its associated network and the primary CIDR range.

UsableSubnetworkSecondaryRange

Secondary IP range of a usable subnetwork.

VerticalPodAutoscaling

VerticalPodAutoscaling contains global, per-cluster information required by Vertical Pod Autoscaler to automatically adjust the resources of pods controlled by it.

VirtualNic

Configuration of gVNIC feature.

WindowsNodeConfig

Parameters that can be configured on Windows nodes. Windows Node Config that define the parameters that will be used to configure the Windows node pool settings

WorkloadIdentityConfig

Configuration for the use of Kubernetes Service Accounts in GCP IAM policies.

WorkloadMetadataConfig

WorkloadMetadataConfig defines the metadata configuration to expose to workloads on the node pool.

WorkloadPolicyConfig

WorkloadPolicyConfig is the configuration of workload policy for autopilot clusters.

Enums

DatapathProvider

The datapath provider selects the implementation of the Kubernetes networking model for service resolution and network policy enforcement.

IPv6AccessType

Possible values for IPv6 access type

InTransitEncryptionConfig

Options for in-transit encryption.

NodePoolUpdateStrategy

Strategy used for node pool update.

PrivateIPv6GoogleAccess

PrivateIPv6GoogleAccess controls whether and how the pods can communicate with Google Services through gRPC over IPv6.

StackType

Possible values for IP stack type

UpgradeResourceType

UpgradeResourceType is the resource type that is upgrading. It is used in upgrade notifications.