Struct google_api_proto::google::container::v1alpha1::Cluster

pub struct Cluster {

    pub name: String,
    pub description: String,
    pub initial_node_count: i32,
    pub node_config: Option<NodeConfig>,
    pub master_auth: Option<MasterAuth>,
    pub logging_service: String,
    pub monitoring_service: String,
    pub network: String,
    pub cluster_ipv4_cidr: String,
    pub addons_config: Option<AddonsConfig>,
    pub subnetwork: String,
    pub node_pools: Vec<NodePool>,
    pub locations: Vec<String>,
    pub enable_kubernetes_alpha: bool,
    pub network_policy: Option<NetworkPolicy>,
    pub ip_allocation_policy: Option<IpAllocationPolicy>,
    pub master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub pod_security_policy_config: Option<PodSecurityPolicyConfig>,
    pub self_link: String,
    pub zone: String,
    pub endpoint: String,
    pub initial_cluster_version: String,
    pub current_master_version: String,
    pub current_node_version: String,
    pub create_time: String,
    pub status: i32,
    pub status_message: String,
    pub node_ipv4_cidr_size: i32,
    pub services_ipv4_cidr: String,
    pub instance_group_urls: Vec<String>,
    pub current_node_count: i32,
    pub expire_time: String,
    pub location: String,

}

A Google Container Engine cluster.

Fields
name: String

The name of this cluster. The name must be unique within this project and zone, and can be up to 40 characters with the following restrictions:

Lowercase letters, numbers, and hyphens only.

Must start with a letter.

Must end with a number or a letter.

description: String

An optional description of this cluster.

initial_node_count: i32

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “node_config”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time.

node_config: Option<NodeConfig>

Parameters used in creating the cluster’s nodes. See nodeConfig for the description of its properties. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “initial_node_count”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time. For responses, this field will be populated with the node configuration of the first node pool.

If unspecified, the defaults are used.

master_auth: Option<MasterAuth>

The authentication information for accessing the master endpoint.

logging_service: String

The logging service the cluster should use to write logs. Currently available options:

logging.googleapis.com - the Google Cloud Logging service.

none - no logs will be exported from the cluster.

if left as an empty string,logging.googleapis.com will be used.

monitoring_service: String

The monitoring service the cluster should use to write metrics. Currently available options:

monitoring.googleapis.com - the Google Cloud Monitoring service.

none - no metrics will be exported from the cluster.

if left as an empty string, monitoring.googleapis.com will be used.

network: String

The name of the Google Compute Engine network to which the cluster is connected. If left unspecified, the default network will be used.

cluster_ipv4_cidr: String

The IP address range of the container pods in this cluster, in CIDR notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

addons_config: Option<AddonsConfig>

Configurations for the various addons available to run in the cluster.

subnetwork: String

The name of the Google Compute Engine subnetwork to which the cluster is connected.

node_pools: Vec<NodePool>

The node pools associated with this cluster. This field should not be set if “node_config” or “initial_node_count” are specified.

locations: Vec<String>

The list of Google Compute Engine locations in which the cluster’s nodes should be located.

enable_kubernetes_alpha: bool

Kubernetes alpha features are enabled on this cluster. This includes alpha API groups (e.g. v1alpha1) and features that may not be production ready in the kubernetes version of the master and nodes. The cluster has no SLA for uptime and master/node upgrades are disabled. Alpha enabled clusters are automatically deleted thirty days after creation.

network_policy: Option<NetworkPolicy>

Configuration options for the NetworkPolicy feature.

ip_allocation_policy: Option<IpAllocationPolicy>

Configuration for cluster IP allocation.

master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>

The configuration options for master authorized networks feature.

maintenance_policy: Option<MaintenancePolicy>

Configure the maintenance policy for this cluster.

pod_security_policy_config: Option<PodSecurityPolicyConfig>

Configuration for the PodSecurityPolicy feature.

self_link: String

[Output only] Server-defined URL for the resource.

zone: String

[Output only] The name of the Google Compute Engine zone in which the cluster resides. This field is deprecated, use location instead.

endpoint: String

[Output only] The IP address of this cluster’s master endpoint. The endpoint can be accessed from the internet at <https://username:password@endpoint/.>

See the masterAuth property of this resource for username and password information.

initial_cluster_version: String

The initial Kubernetes version for this cluster. Valid versions are those found in validMasterVersions returned by getServerConfig. The version can be upgraded over time; such upgrades are reflected in currentMasterVersion and currentNodeVersion.

current_master_version: String

[Output only] The current software version of the master endpoint.

current_node_version: String

[Output only] The current version of the node software components. If they are currently at multiple versions because they’re in the process of being upgraded, this reflects the minimum version of all nodes.

create_time: String

[Output only] The time the cluster was created, in RFC3339 text format.

status: i32

[Output only] The current status of this cluster.

status_message: String

[Output only] Additional information about the current status of this cluster, if available.

node_ipv4_cidr_size: i32

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range.

services_ipv4_cidr: String

[Output only] The IP address range of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.

instance_group_urls: Vec<String>

[Output only] The resource URLs of instance groups associated with this cluster.

current_node_count: i32

[Output only] The number of nodes currently in the cluster.

expire_time: String

[Output only] The time the cluster will be automatically deleted in RFC3339 text format.

location: String

[Output only] The name of the Google Compute Engine zone or region in which the cluster resides.

Implementations

impl Cluster

pub fn status(&self) -> Status

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: Status)

Sets status to the provided enum value.