Module google_api_proto::google::container::v1beta1::blue_green_settings

Nested message and enum types in BlueGreenSettings.

Modules

standard_rollout_policy

Nested message and enum types in StandardRolloutPolicy.

Structs

AutoscaledRolloutPolicy

Autoscaled rollout policy uses cluster autoscaler during blue-green upgrades to scale both the green and blue pools.

StandardRolloutPolicy

Standard rollout policy is the default policy for blue-green.

Enums

RolloutPolicy

The rollout policy controls the general rollout progress of blue-green.