Struct google_api_proto::google::container::v1beta1::Cluster

pub struct Cluster {

    pub name: String,
    pub description: String,
    pub initial_node_count: i32,
    pub node_config: Option<NodeConfig>,
    pub master_auth: Option<MasterAuth>,
    pub logging_service: String,
    pub monitoring_service: String,
    pub network: String,
    pub cluster_ipv4_cidr: String,
    pub addons_config: Option<AddonsConfig>,
    pub subnetwork: String,
    pub node_pools: Vec<NodePool>,
    pub locations: Vec<String>,
    pub enable_kubernetes_alpha: bool,
    pub enable_k8s_beta_apis: Option<K8sBetaApiConfig>,
    pub resource_labels: BTreeMap<String, String>,
    pub label_fingerprint: String,
    pub legacy_abac: Option<LegacyAbac>,
    pub network_policy: Option<NetworkPolicy>,
    pub ip_allocation_policy: Option<IpAllocationPolicy>,
    pub master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub binary_authorization: Option<BinaryAuthorization>,
    pub pod_security_policy_config: Option<PodSecurityPolicyConfig>,
    pub autoscaling: Option<ClusterAutoscaling>,
    pub network_config: Option<NetworkConfig>,
    pub private_cluster: bool,
    pub master_ipv4_cidr_block: String,
    pub default_max_pods_constraint: Option<MaxPodsConstraint>,
    pub resource_usage_export_config: Option<ResourceUsageExportConfig>,
    pub authenticator_groups_config: Option<AuthenticatorGroupsConfig>,
    pub private_cluster_config: Option<PrivateClusterConfig>,
    pub vertical_pod_autoscaling: Option<VerticalPodAutoscaling>,
    pub shielded_nodes: Option<ShieldedNodes>,
    pub release_channel: Option<ReleaseChannel>,
    pub workload_identity_config: Option<WorkloadIdentityConfig>,
    pub workload_certificates: Option<WorkloadCertificates>,
    pub mesh_certificates: Option<MeshCertificates>,
    pub workload_alts_config: Option<WorkloadAltsConfig>,
    pub cost_management_config: Option<CostManagementConfig>,
    pub cluster_telemetry: Option<ClusterTelemetry>,
    pub tpu_config: Option<TpuConfig>,
    pub notification_config: Option<NotificationConfig>,
    pub confidential_nodes: Option<ConfidentialNodes>,
    pub identity_service_config: Option<IdentityServiceConfig>,
    pub self_link: String,
    pub zone: String,
    pub endpoint: String,
    pub initial_cluster_version: String,
    pub current_master_version: String,
    pub current_node_version: String,
    pub create_time: String,
    pub status: i32,
    pub status_message: String,
    pub node_ipv4_cidr_size: i32,
    pub services_ipv4_cidr: String,
    pub instance_group_urls: Vec<String>,
    pub current_node_count: i32,
    pub expire_time: String,
    pub location: String,
    pub enable_tpu: bool,
    pub tpu_ipv4_cidr_block: String,
    pub database_encryption: Option<DatabaseEncryption>,
    pub conditions: Vec<StatusCondition>,
    pub master: Option<Master>,
    pub autopilot: Option<Autopilot>,
    pub id: String,
    pub node_pool_defaults: Option<NodePoolDefaults>,
    pub logging_config: Option<LoggingConfig>,
    pub monitoring_config: Option<MonitoringConfig>,
    pub node_pool_auto_config: Option<NodePoolAutoConfig>,
    pub protect_config: Option<ProtectConfig>,
    pub etag: String,
    pub fleet: Option<Fleet>,
    pub security_posture_config: Option<SecurityPostureConfig>,
    pub enterprise_config: Option<EnterpriseConfig>,
    pub secret_manager_config: Option<SecretManagerConfig>,
    pub compliance_posture_config: Option<CompliancePostureConfig>,
    pub satisfies_pzs: Option<bool>,
    pub satisfies_pzi: Option<bool>,

}

A Google Kubernetes Engine cluster.

Fields
name: String

The name of this cluster. The name must be unique within this project and location (e.g. zone or region), and can be up to 40 characters with the following restrictions:

Lowercase letters, numbers, and hyphens only.

Must start with a letter.

Must end with a number or a letter.

description: String

An optional description of this cluster.

initial_node_count: i32
👎Deprecated

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “node_config”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time.

This field is deprecated, use node_pool.initial_node_count instead.

node_config: Option<NodeConfig>
👎Deprecated

Parameters used in creating the cluster’s nodes. For requests, this field should only be used in lieu of a “node_pool” object, since this configuration (along with the “initial_node_count”) will be used to create a “NodePool” object with an auto-generated name. Do not use this and a node_pool at the same time. For responses, this field will be populated with the node configuration of the first node pool. (For configuration of each node pool, see node_pool.config)

If unspecified, the defaults are used. This field is deprecated, use node_pool.config instead.

master_auth: Option<MasterAuth>

The authentication information for accessing the master endpoint. If unspecified, the defaults are used: For clusters before v1.12, if master_auth is unspecified, username will be set to “admin”, a random password will be generated, and a client certificate will be issued.

logging_service: String

The logging service the cluster should use to write logs. Currently available options:

logging.googleapis.com/kubernetes - The Cloud Logging service with a Kubernetes-native resource model

logging.googleapis.com - The legacy Cloud Logging service (no longer available as of GKE 1.15).

none - no logs will be exported from the cluster.

If left as an empty string,logging.googleapis.com/kubernetes will be used for GKE 1.14+ or logging.googleapis.com for earlier versions.

monitoring_service: String

The monitoring service the cluster should use to write metrics. Currently available options:

“monitoring.googleapis.com/kubernetes” - The Cloud Monitoring service with a Kubernetes-native resource model

monitoring.googleapis.com - The legacy Cloud Monitoring service (no longer available as of GKE 1.15).

none - No metrics will be exported from the cluster.

If left as an empty string,monitoring.googleapis.com/kubernetes will be used for GKE 1.14+ or monitoring.googleapis.com for earlier versions.

network: String

The name of the Google Compute Engine network to which the cluster is connected. If left unspecified, the default network will be used. On output this shows the network ID instead of the name.

cluster_ipv4_cidr: String

The IP address range of the container pods in this cluster, in CIDR notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

addons_config: Option<AddonsConfig>

Configurations for the various addons available to run in the cluster.

subnetwork: String

The name of the Google Compute Engine subnetwork to which the cluster is connected. On output this shows the subnetwork ID instead of the name.

node_pools: Vec<NodePool>

The node pools associated with this cluster. This field should not be set if “node_config” or “initial_node_count” are specified.

locations: Vec<String>

The list of Google Compute Engine zones in which the cluster’s nodes should be located.

This field provides a default value if NodePool.Locations are not specified during node pool creation.

Warning: changing cluster locations will update the NodePool.Locations of all node pools and will result in nodes being added and/or removed.

enable_kubernetes_alpha: bool

Kubernetes alpha features are enabled on this cluster. This includes alpha API groups (e.g. v1beta1) and features that may not be production ready in the kubernetes version of the master and nodes. The cluster has no SLA for uptime and master/node upgrades are disabled. Alpha enabled clusters are automatically deleted thirty days after creation.

enable_k8s_beta_apis: Option<K8sBetaApiConfig>

Kubernetes open source beta apis enabled on the cluster. Only beta apis.

resource_labels: BTreeMap<String, String>

The resource labels for the cluster to use to annotate any related Google Compute Engine resources.

label_fingerprint: String

The fingerprint of the set of labels for this cluster.

legacy_abac: Option<LegacyAbac>

Configuration for the legacy ABAC authorization mode.

network_policy: Option<NetworkPolicy>

Configuration options for the NetworkPolicy feature.

ip_allocation_policy: Option<IpAllocationPolicy>

Configuration for cluster IP allocation.

master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>

The configuration options for master authorized networks feature.

maintenance_policy: Option<MaintenancePolicy>

Configure the maintenance policy for this cluster.

binary_authorization: Option<BinaryAuthorization>

Configuration for Binary Authorization.

pod_security_policy_config: Option<PodSecurityPolicyConfig>

Configuration for the PodSecurityPolicy feature.

autoscaling: Option<ClusterAutoscaling>

Cluster-level autoscaling configuration.

network_config: Option<NetworkConfig>

Configuration for cluster networking.

private_cluster: bool
👎Deprecated

If this is a private cluster setup. Private clusters are clusters that, by default have no external IP addresses on the nodes and where nodes and the master communicate over private IP addresses. This field is deprecated, use private_cluster_config.enable_private_nodes instead.

master_ipv4_cidr_block: String
👎Deprecated

The IP prefix in CIDR notation to use for the hosted master network. This prefix will be used for assigning private IP addresses to the master or set of masters, as well as the ILB VIP. This field is deprecated, use private_cluster_config.master_ipv4_cidr_block instead.

default_max_pods_constraint: Option<MaxPodsConstraint>

The default constraint on the maximum number of pods that can be run simultaneously on a node in the node pool of this cluster. Only honored if cluster created with IP Alias support.

resource_usage_export_config: Option<ResourceUsageExportConfig>

Configuration for exporting resource usages. Resource usage export is disabled when this config unspecified.

authenticator_groups_config: Option<AuthenticatorGroupsConfig>

Configuration controlling RBAC group membership information.

private_cluster_config: Option<PrivateClusterConfig>

Configuration for private cluster.

vertical_pod_autoscaling: Option<VerticalPodAutoscaling>

Cluster-level Vertical Pod Autoscaling configuration.

shielded_nodes: Option<ShieldedNodes>

Shielded Nodes configuration.

release_channel: Option<ReleaseChannel>

Release channel configuration. If left unspecified on cluster creation and a version is specified, the cluster is enrolled in the most mature release channel where the version is available (first checking STABLE, then REGULAR, and finally RAPID). Otherwise, if no release channel configuration and no version is specified, the cluster is enrolled in the REGULAR channel with its default version.

workload_identity_config: Option<WorkloadIdentityConfig>

Configuration for the use of Kubernetes Service Accounts in GCP IAM policies.

workload_certificates: Option<WorkloadCertificates>

Configuration for issuance of mTLS keys and certificates to Kubernetes pods.

mesh_certificates: Option<MeshCertificates>

Configuration for issuance of mTLS keys and certificates to Kubernetes pods.

workload_alts_config: Option<WorkloadAltsConfig>

Configuration for direct-path (via ALTS) with workload identity.

cost_management_config: Option<CostManagementConfig>

Configuration for the fine-grained cost management feature.

cluster_telemetry: Option<ClusterTelemetry>

Telemetry integration for the cluster.

tpu_config: Option<TpuConfig>

Configuration for Cloud TPU support;

notification_config: Option<NotificationConfig>

Notification configuration of the cluster.

confidential_nodes: Option<ConfidentialNodes>

Configuration of Confidential Nodes. All the nodes in the cluster will be Confidential VM once enabled.

identity_service_config: Option<IdentityServiceConfig>

Configuration for Identity Service component.

self_link: String

[Output only] Server-defined URL for the resource.

zone: String
👎Deprecated

[Output only] The name of the Google Compute Engine zone in which the cluster resides. This field is deprecated, use location instead.

endpoint: String

[Output only] The IP address of this cluster’s master endpoint. The endpoint can be accessed from the internet at <https://username:password@endpoint/.>

See the masterAuth property of this resource for username and password information.

initial_cluster_version: String

The initial Kubernetes version for this cluster. Valid versions are those found in validMasterVersions returned by getServerConfig. The version can be upgraded over time; such upgrades are reflected in currentMasterVersion and currentNodeVersion.

Users may specify either explicit versions offered by Kubernetes Engine or version aliases, which have the following behavior:

“latest”: picks the highest valid Kubernetes version

“1.X”: picks the highest valid patch+gke.N patch in the 1.X version

“1.X.Y”: picks the highest valid gke.N patch in the 1.X.Y version

“1.X.Y-gke.N”: picks an explicit Kubernetes version

“”,“-”: picks the default Kubernetes version

current_master_version: String

[Output only] The current software version of the master endpoint.

current_node_version: String
👎Deprecated

[Output only] Deprecated, use NodePool.version instead. The current version of the node software components. If they are currently at multiple versions because they’re in the process of being upgraded, this reflects the minimum version of all nodes.

create_time: String

[Output only] The time the cluster was created, in RFC3339 text format.

status: i32

[Output only] The current status of this cluster.

status_message: String
👎Deprecated

[Output only] Deprecated. Use conditions instead. Additional information about the current status of this cluster, if available.

node_ipv4_cidr_size: i32

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range. This field will only be set when cluster is in route-based network mode.

services_ipv4_cidr: String

[Output only] The IP address range of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.

instance_group_urls: Vec<String>
👎Deprecated

Deprecated. Use node_pools.instance_group_urls.

current_node_count: i32
👎Deprecated

[Output only] The number of nodes currently in the cluster. Deprecated. Call Kubernetes API directly to retrieve node information.

expire_time: String

[Output only] The time the cluster will be automatically deleted in RFC3339 text format.

location: String

[Output only] The name of the Google Compute Engine zone or region in which the cluster resides.

enable_tpu: bool

Enable the ability to use Cloud TPUs in this cluster. This field is deprecated, use tpu_config.enabled instead.

tpu_ipv4_cidr_block: String

[Output only] The IP address range of the Cloud TPUs in this cluster, in CIDR notation (e.g. 1.2.3.4/29).

database_encryption: Option<DatabaseEncryption>

Configuration of etcd encryption.

conditions: Vec<StatusCondition>

Which conditions caused the current cluster state.

master: Option<Master>

Configuration for master components.

autopilot: Option<Autopilot>

Autopilot configuration for the cluster.

id: String

Output only. Unique id for the cluster.

node_pool_defaults: Option<NodePoolDefaults>

Default NodePool settings for the entire cluster. These settings are overridden if specified on the specific NodePool object.

logging_config: Option<LoggingConfig>

Logging configuration for the cluster.

monitoring_config: Option<MonitoringConfig>

Monitoring configuration for the cluster.

node_pool_auto_config: Option<NodePoolAutoConfig>

Node pool configs that apply to all auto-provisioned node pools in autopilot clusters and node auto-provisioning enabled clusters.

protect_config: Option<ProtectConfig>
👎Deprecated

Deprecated: Use SecurityPostureConfig instead. Enable/Disable Protect API features for the cluster.

etag: String

This checksum is computed by the server based on the value of cluster fields, and may be sent on update requests to ensure the client has an up-to-date value before proceeding.

fleet: Option<Fleet>

Fleet information for the cluster.

security_posture_config: Option<SecurityPostureConfig>

Enable/Disable Security Posture API features for the cluster.

enterprise_config: Option<EnterpriseConfig>

GKE Enterprise Configuration.

secret_manager_config: Option<SecretManagerConfig>

Secret CSI driver configuration.

compliance_posture_config: Option<CompliancePostureConfig>

Enable/Disable Compliance Posture features for the cluster.

satisfies_pzs: Option<bool>

Output only. Reserved for future use.

satisfies_pzi: Option<bool>

Output only. Reserved for future use.

Implementations

impl Cluster

pub fn status(&self) -> Status

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: Status)

Sets status to the provided enum value.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.

pub fn satisfies_pzi(&self) -> bool

Returns the value of satisfies_pzi, or the default value if satisfies_pzi is unset.