Struct google_api_proto::google::container::v1beta1::IpAllocationPolicy

pub struct IpAllocationPolicy {

    pub use_ip_aliases: bool,
    pub create_subnetwork: bool,
    pub subnetwork_name: String,
    pub cluster_ipv4_cidr: String,
    pub node_ipv4_cidr: String,
    pub services_ipv4_cidr: String,
    pub cluster_secondary_range_name: String,
    pub services_secondary_range_name: String,
    pub cluster_ipv4_cidr_block: String,
    pub node_ipv4_cidr_block: String,
    pub services_ipv4_cidr_block: String,
    pub allow_route_overlap: bool,
    pub tpu_ipv4_cidr_block: String,
    pub use_routes: bool,
    pub stack_type: i32,
    pub ipv6_access_type: i32,
    pub pod_cidr_overprovision_config: Option<PodCidrOverprovisionConfig>,
    pub subnet_ipv6_cidr_block: String,
    pub services_ipv6_cidr_block: String,
    pub additional_pod_ranges_config: Option<AdditionalPodRangesConfig>,
    pub default_pod_ipv4_range_utilization: f64,

}

Configuration for controlling how IPs are allocated in the cluster.

Fields
use_ip_aliases: bool

Whether alias IPs will be used for pod IPs in the cluster. This is used in conjunction with use_routes. It cannot be true if use_routes is true. If both use_ip_aliases and use_routes are false, then the server picks the default IP allocation mode

create_subnetwork: bool

Whether a new subnetwork will be created automatically for the cluster.

This field is only applicable when use_ip_aliases is true.

subnetwork_name: String

A custom subnetwork name to be used if create_subnetwork is true. If this field is empty, then an automatic name will be chosen for the new subnetwork.

cluster_ipv4_cidr: String
👎Deprecated

This field is deprecated, use cluster_ipv4_cidr_block.

node_ipv4_cidr: String
👎Deprecated

This field is deprecated, use node_ipv4_cidr_block.

services_ipv4_cidr: String
👎Deprecated

This field is deprecated, use services_ipv4_cidr_block.

cluster_secondary_range_name: String

The name of the secondary range to be used for the cluster CIDR block. The secondary range will be used for pod IP addresses. This must be an existing secondary range associated with the cluster subnetwork.

This field is only applicable with use_ip_aliases and create_subnetwork is false.

services_secondary_range_name: String

The name of the secondary range to be used as for the services CIDR block. The secondary range will be used for service ClusterIPs. This must be an existing secondary range associated with the cluster subnetwork.

This field is only applicable with use_ip_aliases and create_subnetwork is false.

cluster_ipv4_cidr_block: String

The IP address range for the cluster pod IPs. If this field is set, then cluster.cluster_ipv4_cidr must be left blank.

This field is only applicable when use_ip_aliases is true.

Set to blank to have a range chosen with the default size.

Set to /netmask (e.g. /14) to have a range chosen with a specific netmask.

Set to a CIDR notation (e.g. 10.96.0.0/14) from the RFC-1918 private networks (e.g. 10.0.0.0/8, 172.16.0.0/12, 192.168.0.0/16) to pick a specific range to use.

node_ipv4_cidr_block: String

The IP address range of the instance IPs in this cluster.

This is applicable only if create_subnetwork is true.

Set to blank to have a range chosen with the default size.

Set to /netmask (e.g. /14) to have a range chosen with a specific netmask.

Set to a CIDR notation (e.g. 10.96.0.0/14) from the RFC-1918 private networks (e.g. 10.0.0.0/8, 172.16.0.0/12, 192.168.0.0/16) to pick a specific range to use.

services_ipv4_cidr_block: String

The IP address range of the services IPs in this cluster. If blank, a range will be automatically chosen with the default size.

This field is only applicable when use_ip_aliases is true.

Set to blank to have a range chosen with the default size.

Set to /netmask (e.g. /14) to have a range chosen with a specific netmask.

Set to a CIDR notation (e.g. 10.96.0.0/14) from the RFC-1918 private networks (e.g. 10.0.0.0/8, 172.16.0.0/12, 192.168.0.0/16) to pick a specific range to use.

allow_route_overlap: bool

If true, allow allocation of cluster CIDR ranges that overlap with certain kinds of network routes. By default we do not allow cluster CIDR ranges to intersect with any user declared routes. With allow_route_overlap == true, we allow overlapping with CIDR ranges that are larger than the cluster CIDR range.

If this field is set to true, then cluster and services CIDRs must be fully-specified (e.g. 10.96.0.0/14, but not /14), which means:

When use_ip_aliases is true, cluster_ipv4_cidr_block and services_ipv4_cidr_block must be fully-specified.

When use_ip_aliases is false, cluster.cluster_ipv4_cidr muse be fully-specified.

tpu_ipv4_cidr_block: String

The IP address range of the Cloud TPUs in this cluster. If unspecified, a range will be automatically chosen with the default size.

This field is only applicable when use_ip_aliases is true.

If unspecified, the range will use the default size.

Set to /netmask (e.g. /14) to have a range chosen with a specific netmask.

Set to a CIDR notation (e.g. 10.96.0.0/14) from the RFC-1918 private networks (e.g. 10.0.0.0/8, 172.16.0.0/12, 192.168.0.0/16) to pick a specific range to use. This field is deprecated, use cluster.tpu_config.ipv4_cidr_block instead.

use_routes: bool

Whether routes will be used for pod IPs in the cluster. This is used in conjunction with use_ip_aliases. It cannot be true if use_ip_aliases is true. If both use_ip_aliases and use_routes are false, then the server picks the default IP allocation mode

stack_type: i32

IP stack type

ipv6_access_type: i32

The ipv6 access type (internal or external) when create_subnetwork is true

pod_cidr_overprovision_config: Option<PodCidrOverprovisionConfig>

[PRIVATE FIELD] Pod CIDR size overprovisioning config for the cluster.

Pod CIDR size per node depends on max_pods_per_node. By default, the value of max_pods_per_node is doubled and then rounded off to next power of 2 to get the size of pod CIDR block per node. Example: max_pods_per_node of 30 would result in 64 IPs (/26).

This config can disable the doubling of IPs (we still round off to next power of 2) Example: max_pods_per_node of 30 will result in 32 IPs (/27) when overprovisioning is disabled.

subnet_ipv6_cidr_block: String

Output only. [Output only] The subnet’s IPv6 CIDR block used by nodes and pods.

services_ipv6_cidr_block: String

Output only. [Output only] The services IPv6 CIDR block for the cluster.

additional_pod_ranges_config: Option<AdditionalPodRangesConfig>

Output only. [Output only] The additional pod ranges that are added to the cluster. These pod ranges can be used by new node pools to allocate pod IPs automatically. Once the range is removed it will not show up in IPAllocationPolicy.

default_pod_ipv4_range_utilization: f64

Output only. [Output only] The utilization of the cluster default IPv4 range for the pod. The ratio is Usage/[Total number of IPs in the secondary range], Usage=numNodesnumZonespodIPsPerNode.

Implementations

impl IpAllocationPolicy

pub fn stack_type(&self) -> StackType

Returns the enum value of stack_type, or the default if the field is set to an invalid enum value.

pub fn set_stack_type(&mut self, value: StackType)

Sets stack_type to the provided enum value.

pub fn ipv6_access_type(&self) -> IPv6AccessType

Returns the enum value of ipv6_access_type, or the default if the field is set to an invalid enum value.

pub fn set_ipv6_access_type(&mut self, value: IPv6AccessType)

Sets ipv6_access_type to the provided enum value.