Struct google_api_proto::google::dataflow::v1beta3::jobs_v1_beta3_client::JobsV1Beta3Client

pub struct JobsV1Beta3Client<T> { /* private fields */ }

Provides a method to create and modify Google Cloud Dataflow jobs. A Job is a multi-stage computation graph run by the Cloud Dataflow service.

Implementations

impl<T> JobsV1Beta3Client<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> JobsV1Beta3Client<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_job( &mut self, request: impl IntoRequest<CreateJobRequest> ) -> Result<Response<Job>, Status>

Creates a Cloud Dataflow job.

To create a job, we recommend using projects.locations.jobs.create with a [regional endpoint] (https://cloud.google.com/dataflow/docs/concepts/regional-endpoints). Using projects.jobs.create is not recommended, as your job will always start in us-central1.

pub async fn get_job( &mut self, request: impl IntoRequest<GetJobRequest> ) -> Result<Response<Job>, Status>

Gets the state of the specified Cloud Dataflow job.

To get the state of a job, we recommend using projects.locations.jobs.get with a [regional endpoint] (https://cloud.google.com/dataflow/docs/concepts/regional-endpoints). Using projects.jobs.get is not recommended, as you can only get the state of jobs that are running in us-central1.

pub async fn update_job( &mut self, request: impl IntoRequest<UpdateJobRequest> ) -> Result<Response<Job>, Status>

Updates the state of an existing Cloud Dataflow job.

To update the state of an existing job, we recommend using projects.locations.jobs.update with a [regional endpoint] (https://cloud.google.com/dataflow/docs/concepts/regional-endpoints). Using projects.jobs.update is not recommended, as you can only update the state of jobs that are running in us-central1.

pub async fn list_jobs( &mut self, request: impl IntoRequest<ListJobsRequest> ) -> Result<Response<ListJobsResponse>, Status>

List the jobs of a project.

To list the jobs of a project in a region, we recommend using projects.locations.jobs.list with a [regional endpoint] (https://cloud.google.com/dataflow/docs/concepts/regional-endpoints). To list the all jobs across all regions, use projects.jobs.aggregated. Using projects.jobs.list is not recommended, as you can only get the list of jobs that are running in us-central1.

pub async fn aggregated_list_jobs( &mut self, request: impl IntoRequest<ListJobsRequest> ) -> Result<Response<ListJobsResponse>, Status>

List the jobs of a project across all regions.

pub async fn check_active_jobs( &mut self, request: impl IntoRequest<CheckActiveJobsRequest> ) -> Result<Response<CheckActiveJobsResponse>, Status>

Check for existence of active jobs in the given project across all regions.

pub async fn snapshot_job( &mut self, request: impl IntoRequest<SnapshotJobRequest> ) -> Result<Response<Snapshot>, Status>

Snapshot the state of a streaming job.