Module google_api_proto::google::datastore::admin::v1

Modules

common_metadata

Nested message and enum types in CommonMetadata.

datastore_admin_client

Generated client implementations.

index

Nested message and enum types in Index.

migration_progress_event

Nested message and enum types in MigrationProgressEvent.

Structs

CommonMetadata

Metadata common to all Datastore Admin operations.

CreateIndexRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.CreateIndex][google.datastore.admin.v1.DatastoreAdmin.CreateIndex].

DatastoreFirestoreMigrationMetadata

Metadata for Datastore to Firestore migration operations.

DeleteIndexRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.DeleteIndex][google.datastore.admin.v1.DatastoreAdmin.DeleteIndex].

EntityFilter

Identifies a subset of entities in a project. This is specified as combinations of kinds and namespaces (either or both of which may be all, as described in the following examples). Example usage:

ExportEntitiesMetadata

Metadata for ExportEntities operations.

ExportEntitiesRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.ExportEntities][google.datastore.admin.v1.DatastoreAdmin.ExportEntities].

ExportEntitiesResponse

The response for [google.datastore.admin.v1.DatastoreAdmin.ExportEntities][google.datastore.admin.v1.DatastoreAdmin.ExportEntities].

GetIndexRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.GetIndex][google.datastore.admin.v1.DatastoreAdmin.GetIndex].

ImportEntitiesMetadata

Metadata for ImportEntities operations.

ImportEntitiesRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.ImportEntities][google.datastore.admin.v1.DatastoreAdmin.ImportEntities].

Index

Datastore composite index definition.

IndexOperationMetadata

Metadata for Index operations.

ListIndexesRequest

The request for [google.datastore.admin.v1.DatastoreAdmin.ListIndexes][google.datastore.admin.v1.DatastoreAdmin.ListIndexes].

ListIndexesResponse

The response for [google.datastore.admin.v1.DatastoreAdmin.ListIndexes][google.datastore.admin.v1.DatastoreAdmin.ListIndexes].

MigrationProgressEvent

An event signifying the start of a new step in a migration from Cloud Datastore to Cloud Firestore in Datastore mode.

MigrationStateEvent

An event signifying a change in state of a migration from Cloud Datastore to Cloud Firestore in Datastore mode.

Progress

Measures the progress of a particular metric.

Enums

MigrationState

States for a migration.

MigrationStep

Steps in a migration.

OperationType

Operation types.