Module google_api_proto::google::datastore::v1

Modules

aggregation_query

Nested message and enum types in AggregationQuery.

commit_request

Nested message and enum types in CommitRequest.

composite_filter

Nested message and enum types in CompositeFilter.

datastore_client

Generated client implementations.

entity_result

Nested message and enum types in EntityResult.

filter

Nested message and enum types in Filter.

gql_query_parameter

Nested message and enum types in GqlQueryParameter.

key

Nested message and enum types in Key.

mutation

Nested message and enum types in Mutation.

property_filter

Nested message and enum types in PropertyFilter.

property_order

Nested message and enum types in PropertyOrder.

query_result_batch

Nested message and enum types in QueryResultBatch.

read_options

Nested message and enum types in ReadOptions.

run_aggregation_query_request

Nested message and enum types in RunAggregationQueryRequest.

run_query_request

Nested message and enum types in RunQueryRequest.

transaction_options

Nested message and enum types in TransactionOptions.

value

Nested message and enum types in Value.

Structs

AggregationQuery

Datastore query for running an aggregation over a [Query][google.datastore.v1.Query].

AggregationResult

The result of a single bucket from a Datastore aggregation query.

AggregationResultBatch

A batch of aggregation results produced by an aggregation query.

AllocateIdsRequest

The request for [Datastore.AllocateIds][google.datastore.v1.Datastore.AllocateIds].

AllocateIdsResponse

The response for [Datastore.AllocateIds][google.datastore.v1.Datastore.AllocateIds].

ArrayValue

An array value.

BeginTransactionRequest

The request for [Datastore.BeginTransaction][google.datastore.v1.Datastore.BeginTransaction].

BeginTransactionResponse

The response for [Datastore.BeginTransaction][google.datastore.v1.Datastore.BeginTransaction].

CommitRequest

The request for [Datastore.Commit][google.datastore.v1.Datastore.Commit].

CommitResponse

The response for [Datastore.Commit][google.datastore.v1.Datastore.Commit].

CompositeFilter

A filter that merges multiple other filters using the given operator.

Entity

A Datastore data object.

EntityResult

The result of fetching an entity from Datastore.

ExecutionStats

Execution statistics for the query.

ExplainMetrics

Explain metrics for the query.

ExplainOptions

Explain options for the query.

Filter

A holder for any type of filter.

GqlQuery

A GQL query.

GqlQueryParameter

A binding parameter for a GQL query.

Key

A unique identifier for an entity. If a key’s partition ID or any of its path kinds or names are reserved/read-only, the key is reserved/read-only. A reserved/read-only key is forbidden in certain documented contexts.

KindExpression

A representation of a kind.

LookupRequest

The request for [Datastore.Lookup][google.datastore.v1.Datastore.Lookup].

LookupResponse

The response for [Datastore.Lookup][google.datastore.v1.Datastore.Lookup].

Mutation

A mutation to apply to an entity.

MutationResult

The result of applying a mutation.

PartitionId

A partition ID identifies a grouping of entities. The grouping is always by project and namespace, however the namespace ID may be empty.

PlanSummary

Planning phase information for the query.

Projection

A representation of a property in a projection.

PropertyFilter

A filter on a specific property.

PropertyMask

The set of arbitrarily nested property paths used to restrict an operation to only a subset of properties in an entity.

PropertyOrder

The desired order for a specific property.

PropertyReference

A reference to a property relative to the kind expressions.

Query

A query for entities.

QueryResultBatch

A batch of results produced by a query.

ReadOptions

The options shared by read requests.

ReserveIdsRequest

The request for [Datastore.ReserveIds][google.datastore.v1.Datastore.ReserveIds].

ReserveIdsResponse

The response for [Datastore.ReserveIds][google.datastore.v1.Datastore.ReserveIds].

RollbackRequest

The request for [Datastore.Rollback][google.datastore.v1.Datastore.Rollback].

RollbackResponse

The response for [Datastore.Rollback][google.datastore.v1.Datastore.Rollback]. (an empty message).

RunAggregationQueryRequest

The request for [Datastore.RunAggregationQuery][google.datastore.v1.Datastore.RunAggregationQuery].

RunAggregationQueryResponse

The response for [Datastore.RunAggregationQuery][google.datastore.v1.Datastore.RunAggregationQuery].

RunQueryRequest

The request for [Datastore.RunQuery][google.datastore.v1.Datastore.RunQuery].

RunQueryResponse

The response for [Datastore.RunQuery][google.datastore.v1.Datastore.RunQuery].

TransactionOptions

Options for beginning a new transaction.

Value

A message that can hold any of the supported value types and associated metadata.