Struct google_api_proto::google::devtools::artifactregistry::v1::Repository

pub struct Repository {

    pub name: String,
    pub format: i32,
    pub description: String,
    pub labels: BTreeMap<String, String>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub kms_key_name: String,
    pub mode: i32,
    pub cleanup_policies: BTreeMap<String, CleanupPolicy>,
    pub size_bytes: i64,
    pub satisfies_pzs: bool,
    pub cleanup_policy_dry_run: bool,
    pub format_config: Option<FormatConfig>,
    pub mode_config: Option<ModeConfig>,

}

A Repository for storing artifacts with a specific format.

Fields
name: String

The name of the repository, for example: projects/p1/locations/us-central1/repositories/repo1.

format: i32

Optional. The format of packages that are stored in the repository.

description: String

The user-provided description of the repository.

labels: BTreeMap<String, String>

Labels with user-defined metadata. This field may contain up to 64 entries. Label keys and values may be no longer than 63 characters. Label keys must begin with a lowercase letter and may only contain lowercase letters, numeric characters, underscores, and dashes.

create_time: Option<Timestamp>

Output only. The time when the repository was created.

update_time: Option<Timestamp>

Output only. The time when the repository was last updated.

kms_key_name: String

The Cloud KMS resource name of the customer managed encryption key that’s used to encrypt the contents of the Repository. Has the form: projects/my-project/locations/my-region/keyRings/my-kr/cryptoKeys/my-key. This value may not be changed after the Repository has been created.

mode: i32

Optional. The mode of the repository.

cleanup_policies: BTreeMap<String, CleanupPolicy>

Optional. Cleanup policies for this repository. Cleanup policies indicate when certain package versions can be automatically deleted. Map keys are policy IDs supplied by users during policy creation. They must unique within a repository and be under 128 characters in length.

size_bytes: i64

Output only. The size, in bytes, of all artifact storage in this repository. Repositories that are generally available or in public preview use this to calculate storage costs.

satisfies_pzs: bool

Output only. If set, the repository satisfies physical zone separation.

cleanup_policy_dry_run: bool

Optional. If true, the cleanup pipeline is prevented from deleting versions in this repository.

format_config: Option<FormatConfig>

Repository-specific configurations.

mode_config: Option<ModeConfig>

Repository configuration specific to the Mode value being selected (Remote or Virtual)

Implementations

impl Repository

pub fn format(&self) -> Format

Returns the enum value of format, or the default if the field is set to an invalid enum value.

pub fn set_format(&mut self, value: Format)

Sets format to the provided enum value.

pub fn mode(&self) -> Mode

Returns the enum value of mode, or the default if the field is set to an invalid enum value.

pub fn set_mode(&mut self, value: Mode)

Sets mode to the provided enum value.