Struct google_api_proto::google::devtools::artifactregistry::v1beta2::artifact_registry_client::ArtifactRegistryClient

pub struct ArtifactRegistryClient<T> { /* private fields */ }

The Artifact Registry API service.

Artifact Registry is an artifact management system for storing artifacts from different package management systems.

The resources managed by this API are:

Repositories, which group packages and their data.

Packages, which group versions and their tags.

Versions, which are specific forms of a package.

Tags, which represent alternative names for versions.

Files, which contain content and are optionally associated with a Package or Version.

Implementations

impl<T> ArtifactRegistryClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> ArtifactRegistryClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn import_apt_artifacts(
&mut self,
request: impl IntoRequest<ImportAptArtifactsRequest>
) -> Result<Response<Operation>, Status>

Imports Apt artifacts. The returned Operation will complete once the resources are imported. Package, Version, and File resources are created based on the imported artifacts. Imported artifacts that conflict with existing resources are ignored.

pub async fn import_yum_artifacts(
&mut self,
request: impl IntoRequest<ImportYumArtifactsRequest>
) -> Result<Response<Operation>, Status>

Imports Yum (RPM) artifacts. The returned Operation will complete once the resources are imported. Package, Version, and File resources are created based on the imported artifacts. Imported artifacts that conflict with existing resources are ignored.

pub async fn list_repositories(
&mut self,
request: impl IntoRequest<ListRepositoriesRequest>
) -> Result<Response<ListRepositoriesResponse>, Status>

Lists repositories.

pub async fn get_repository(
&mut self,
request: impl IntoRequest<GetRepositoryRequest>
) -> Result<Response<Repository>, Status>

Gets a repository.

pub async fn create_repository(
&mut self,
request: impl IntoRequest<CreateRepositoryRequest>
) -> Result<Response<Operation>, Status>

Creates a repository. The returned Operation will finish once the repository has been created. Its response will be the created Repository.

pub async fn update_repository(
&mut self,
request: impl IntoRequest<UpdateRepositoryRequest>
) -> Result<Response<Repository>, Status>

Updates a repository.

pub async fn delete_repository(
&mut self,
request: impl IntoRequest<DeleteRepositoryRequest>
) -> Result<Response<Operation>, Status>

Deletes a repository and all of its contents. The returned Operation will finish once the repository has been deleted. It will not have any Operation metadata and will return a google.protobuf.Empty response.

pub async fn list_packages(
&mut self,
request: impl IntoRequest<ListPackagesRequest>
) -> Result<Response<ListPackagesResponse>, Status>

Lists packages.

pub async fn get_package(
&mut self,
request: impl IntoRequest<GetPackageRequest>
) -> Result<Response<Package>, Status>

Gets a package.

pub async fn delete_package(
&mut self,
request: impl IntoRequest<DeletePackageRequest>
) -> Result<Response<Operation>, Status>

Deletes a package and all of its versions and tags. The returned operation will complete once the package has been deleted.

pub async fn list_versions(
&mut self,
request: impl IntoRequest<ListVersionsRequest>
) -> Result<Response<ListVersionsResponse>, Status>

Lists versions.

pub async fn get_version(
&mut self,
request: impl IntoRequest<GetVersionRequest>
) -> Result<Response<Version>, Status>

Gets a version

pub async fn delete_version(
&mut self,
request: impl IntoRequest<DeleteVersionRequest>
) -> Result<Response<Operation>, Status>

Deletes a version and all of its content. The returned operation will complete once the version has been deleted.

pub async fn list_files(
&mut self,
request: impl IntoRequest<ListFilesRequest>
) -> Result<Response<ListFilesResponse>, Status>

Lists files.

pub async fn get_file(
&mut self,
request: impl IntoRequest<GetFileRequest>
) -> Result<Response<File>, Status>

Gets a file.

pub async fn list_tags(
&mut self,
request: impl IntoRequest<ListTagsRequest>
) -> Result<Response<ListTagsResponse>, Status>

Lists tags.

pub async fn get_tag(
&mut self,
request: impl IntoRequest<GetTagRequest>
) -> Result<Response<Tag>, Status>

Gets a tag.

pub async fn create_tag(
&mut self,
request: impl IntoRequest<CreateTagRequest>
) -> Result<Response<Tag>, Status>

Creates a tag.

pub async fn update_tag(
&mut self,
request: impl IntoRequest<UpdateTagRequest>
) -> Result<Response<Tag>, Status>

Updates a tag.

pub async fn delete_tag(
&mut self,
request: impl IntoRequest<DeleteTagRequest>
) -> Result<Response<()>, Status>

Deletes a tag.

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Updates the IAM policy for a given resource.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the IAM policy for a given resource.

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Tests if the caller has a list of permissions on a resource.

pub async fn get_project_settings(
&mut self,
request: impl IntoRequest<GetProjectSettingsRequest>
) -> Result<Response<ProjectSettings>, Status>

Retrieves the Settings for the Project.

pub async fn update_project_settings(
&mut self,
request: impl IntoRequest<UpdateProjectSettingsRequest>
) -> Result<Response<ProjectSettings>, Status>

Updates the Settings for the Project.