Module google_api_proto::google::devtools::build::v1::build_event

Nested message and enum types in BuildEvent.

Modules

build_component_stream_finished

Nested message and enum types in BuildComponentStreamFinished.

console_output

Nested message and enum types in ConsoleOutput.

Structs

BuildComponentStreamFinished

Notification of the end of a build event stream published by a build component other than CONTROLLER (See StreamId.BuildComponents).

BuildEnqueued

Notification that the build request is enqueued.

BuildFinished

Notification that the build request has finished, and no further invocations will occur. Note that this applies to the entire Build. Individual invocations trigger InvocationFinished when they finish.

ConsoleOutput

Textual output written to standard output or standard error.

InvocationAttemptFinished

Notification that an invocation attempt has finished.

InvocationAttemptStarted

Notification that the build system has attempted to run the build tool.

Enums

Event

////////////////////////////////////////////////////////////////////////// Events that indicate a state change of a build request in the build queue.