Struct google_api_proto::google::devtools::cloudbuild::v1::BuildOptions

pub struct BuildOptions {

    pub source_provenance_hash: Vec<i32>,
    pub requested_verify_option: i32,
    pub machine_type: i32,
    pub disk_size_gb: i64,
    pub substitution_option: i32,
    pub dynamic_substitutions: bool,
    pub automap_substitutions: bool,
    pub log_streaming_option: i32,
    pub worker_pool: String,
    pub pool: Option<PoolOption>,
    pub logging: i32,
    pub env: Vec<String>,
    pub secret_env: Vec<String>,
    pub volumes: Vec<Volume>,
    pub default_logs_bucket_behavior: i32,

}

Optional arguments to enable specific features of builds.

Fields
source_provenance_hash: Vec<i32>

Requested hash for SourceProvenance.

requested_verify_option: i32

Requested verifiability options.

machine_type: i32

Compute Engine machine type on which to run the build.

disk_size_gb: i64

Requested disk size for the VM that runs the build. Note that this is NOT “disk free”; some of the space will be used by the operating system and build utilities. Also note that this is the minimum disk size that will be allocated for the build – the build may run with a larger disk than requested. At present, the maximum disk size is 2000GB; builds that request more than the maximum are rejected with an error.

substitution_option: i32

Option to specify behavior when there is an error in the substitution checks.

NOTE: this is always set to ALLOW_LOOSE for triggered builds and cannot be overridden in the build configuration file.

dynamic_substitutions: bool

Option to specify whether or not to apply bash style string operations to the substitutions.

NOTE: this is always enabled for triggered builds and cannot be overridden in the build configuration file.

automap_substitutions: bool

Option to include built-in and custom substitutions as env variables for all build steps.

log_streaming_option: i32

Option to define build log streaming behavior to Cloud Storage.

worker_pool: String
👎Deprecated

This field deprecated; please use pool.name instead.

pool: Option<PoolOption>

Optional. Specification for execution on a WorkerPool.

See running builds in a private pool for more information.

logging: i32

Option to specify the logging mode, which determines if and where build logs are stored.

env: Vec<String>

A list of global environment variable definitions that will exist for all build steps in this build. If a variable is defined in both globally and in a build step, the variable will use the build step value.

The elements are of the form “KEY=VALUE” for the environment variable “KEY” being given the value “VALUE”.

secret_env: Vec<String>

A list of global environment variables, which are encrypted using a Cloud Key Management Service crypto key. These values must be specified in the build’s Secret. These variables will be available to all build steps in this build.

volumes: Vec<Volume>

Global list of volumes to mount for ALL build steps

Each volume is created as an empty volume prior to starting the build process. Upon completion of the build, volumes and their contents are discarded. Global volume names and paths cannot conflict with the volumes defined a build step.

Using a global volume in a build with only one step is not valid as it is indicative of a build request with an incorrect configuration.

default_logs_bucket_behavior: i32

Optional. Option to specify how default logs buckets are setup.

Implementations

impl BuildOptions

pub fn source_provenance_hash( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<HashType>>

Returns an iterator which yields the valid enum values contained in source_provenance_hash.

pub fn push_source_provenance_hash(&mut self, value: HashType)

Appends the provided enum value to source_provenance_hash.

pub fn requested_verify_option(&self) -> VerifyOption

Returns the enum value of requested_verify_option, or the default if the field is set to an invalid enum value.

pub fn set_requested_verify_option(&mut self, value: VerifyOption)

Sets requested_verify_option to the provided enum value.

pub fn machine_type(&self) -> MachineType

Returns the enum value of machine_type, or the default if the field is set to an invalid enum value.

pub fn set_machine_type(&mut self, value: MachineType)

Sets machine_type to the provided enum value.

pub fn substitution_option(&self) -> SubstitutionOption

Returns the enum value of substitution_option, or the default if the field is set to an invalid enum value.

pub fn set_substitution_option(&mut self, value: SubstitutionOption)

Sets substitution_option to the provided enum value.

pub fn log_streaming_option(&self) -> LogStreamingOption

Returns the enum value of log_streaming_option, or the default if the field is set to an invalid enum value.

pub fn set_log_streaming_option(&mut self, value: LogStreamingOption)

Sets log_streaming_option to the provided enum value.

pub fn logging(&self) -> LoggingMode

Returns the enum value of logging, or the default if the field is set to an invalid enum value.

pub fn set_logging(&mut self, value: LoggingMode)

Sets logging to the provided enum value.

pub fn default_logs_bucket_behavior(&self) -> DefaultLogsBucketBehavior

Returns the enum value of default_logs_bucket_behavior, or the default if the field is set to an invalid enum value.

pub fn set_default_logs_bucket_behavior( &mut self, value: DefaultLogsBucketBehavior )

Sets default_logs_bucket_behavior to the provided enum value.