Struct google_api_proto::google::devtools::clouddebugger::v2::Variable

pub struct Variable {
    pub name: String,
    pub value: String,
    pub type: String,
    pub members: Vec<Variable>,
    pub var_table_index: Option<i32>,
    pub status: Option<StatusMessage>,
}

Represents a variable or an argument possibly of a compound object type. Note how the following variables are represented:

A simple variable:

int x = 5

{ name: “x”, value: “5”, type: “int” } // Captured variable

A compound object:

struct T { int m1; int m2; }; T x = { 3, 7 };

{ // Captured variable name: “x”, type: “T”, members { name: “m1”, value: “3”, type: “int” }, members { name: “m2”, value: “7”, type: “int” } }

A pointer where the pointee was captured:

T x = { 3, 7 }; T* p = &x;

{ // Captured variable name: “p”, type: “T*”, value: “0x00500500”, members { name: “m1”, value: “3”, type: “int” }, members { name: “m2”, value: “7”, type: “int” } }

A pointer where the pointee was not captured:

T* p = new T;

{ // Captured variable name: “p”, type: “T*”, value: “0x00400400” status { is_error: true, description { format: “unavailable” } } }

The status should describe the reason for the missing value, such as <optimized out>, <inaccessible>, <pointers limit reached>.

Note that a null pointer should not have members.

An unnamed value:

int* p = new int(7);

{ // Captured variable name: “p”, value: “0x00500500”, type: “int*”, members { value: “7”, type: “int” } }

An unnamed pointer where the pointee was not captured:

int* p = new int(7); int** pp = &p;

{ // Captured variable name: “pp”, value: “0x00500500”, type: “int**”, members { value: “0x00400400”, type: “int*” status { is_error: true, description: { format: “unavailable” } } } } }

To optimize computation, memory and network traffic, variables that repeat in the output multiple times can be stored once in a shared variable table and be referenced using the var_table_index field. The variables stored in the shared table are nameless and are essentially a partition of the complete variable. To reconstruct the complete variable, merge the referencing variable with the referenced variable.

When using the shared variable table, the following variables:

 T x = { 3, 7 };
 T* p = &x;
 T& r = x;

 { name: "x", var_table_index: 3, type: "T" }  // Captured variables
 { name: "p", value "0x00500500", type="T*", var_table_index: 3 }
 { name: "r", type="T&", var_table_index: 3 }

 {  // Shared variable table entry #3:
     members { name: "m1", value: "3", type: "int" },
     members { name: "m2", value: "7", type: "int" }
 }

Note that the pointer address is stored with the referencing variable and not with the referenced variable. This allows the referenced variable to be shared between pointers and references.

The type field is optional. The debugger agent may or may not support it.

Fields
name: String

Name of the variable, if any.

value: String

Simple value of the variable.

type: String

Variable type (e.g. MyClass). If the variable is split with var_table_index, type goes next to value. The interpretation of a type is agent specific. It is recommended to include the dynamic type rather than a static type of an object.

members: Vec<Variable>

Members contained or pointed to by the variable.

var_table_index: Option<i32>

Reference to a variable in the shared variable table. More than one variable can reference the same variable in the table. The var_table_index field is an index into variable_table in Breakpoint.

status: Option<StatusMessage>

Status associated with the variable. This field will usually stay unset. A status of a single variable only applies to that variable or expression. The rest of breakpoint data still remains valid. Variables might be reported in error state even when breakpoint is not in final state.

The message may refer to variable name with refers_to set to VARIABLE_NAME. Alternatively refers_to will be set to VARIABLE_VALUE. In either case variable value and members will be unset.

Example of error message applied to name: Invalid expression syntax.

Example of information message applied to value: Not captured.

Examples of error message applied to value:

Malformed string,

Field f not found in class C

Null pointer dereference