Module google_api_proto::google::devtools::cloudtrace::v2::span

Nested message and enum types in Span.

Modules

link

Nested message and enum types in Link.

time_event

Nested message and enum types in TimeEvent.

Structs

Attributes

A set of attributes, each in the format \[KEY]:[VALUE\].

Link

A pointer from the current span to another span in the same trace or in a different trace. For example, this can be used in batching operations, where a single batch handler processes multiple requests from different traces or when the handler receives a request from a different project.

Links

A collection of links, which are references from this span to a span in the same or different trace.

TimeEvent

A time-stamped annotation or message event in the Span.

TimeEvents

A collection of TimeEvents. A TimeEvent is a time-stamped annotation on the span, consisting of either user-supplied key:value pairs, or details of a message sent/received between Spans.

Enums

SpanKind

Type of span. Can be used to specify additional relationships between spans in addition to a parent/child relationship.