Struct google_api_proto::google::devtools::cloudtrace::v2::Span

pub struct Span {

    pub name: String,
    pub span_id: String,
    pub parent_span_id: String,
    pub display_name: Option<TruncatableString>,
    pub start_time: Option<Timestamp>,
    pub end_time: Option<Timestamp>,
    pub attributes: Option<Attributes>,
    pub stack_trace: Option<StackTrace>,
    pub time_events: Option<TimeEvents>,
    pub links: Option<Links>,
    pub status: Option<Status>,
    pub same_process_as_parent_span: Option<bool>,
    pub child_span_count: Option<i32>,
    pub span_kind: i32,

}

A span represents a single operation within a trace. Spans can be nested to form a trace tree. Often, a trace contains a root span that describes the end-to-end latency, and one or more subspans for its sub-operations.

A trace can also contain multiple root spans, or none at all. Spans do not need to be contiguous. There might be gaps or overlaps between spans in a trace.

Fields
name: String

Required. The resource name of the span in the following format:

projects/\[PROJECT_ID\]/traces/\[TRACE_ID\]/spans/\[SPAN_ID\]

\[TRACE_ID\] is a unique identifier for a trace within a project; it is a 32-character hexadecimal encoding of a 16-byte array. It should not be zero.

\[SPAN_ID\] is a unique identifier for a span within a trace; it is a 16-character hexadecimal encoding of an 8-byte array. It should not be zero. .

span_id: String

Required. The \[SPAN_ID\] portion of the span’s resource name.

parent_span_id: String

The \[SPAN_ID\] of this span’s parent span. If this is a root span, then this field must be empty.

display_name: Option<TruncatableString>

Required. A description of the span’s operation (up to 128 bytes). Cloud Trace displays the description in the Cloud console. For example, the display name can be a qualified method name or a file name and a line number where the operation is called. A best practice is to use the same display name within an application and at the same call point. This makes it easier to correlate spans in different traces.

start_time: Option<Timestamp>

Required. The start time of the span. On the client side, this is the time kept by the local machine where the span execution starts. On the server side, this is the time when the server’s application handler starts running.

end_time: Option<Timestamp>

Required. The end time of the span. On the client side, this is the time kept by the local machine where the span execution ends. On the server side, this is the time when the server application handler stops running.

attributes: Option<Attributes>

A set of attributes on the span. You can have up to 32 attributes per span.

stack_trace: Option<StackTrace>

Stack trace captured at the start of the span.

time_events: Option<TimeEvents>

A set of time events. You can have up to 32 annotations and 128 message events per span.

links: Option<Links>

Links associated with the span. You can have up to 128 links per Span.

status: Option<Status>

Optional. The final status for this span.

same_process_as_parent_span: Option<bool>

Optional. Set this parameter to indicate whether this span is in the same process as its parent. If you do not set this parameter, Trace is unable to take advantage of this helpful information.

child_span_count: Option<i32>

Optional. The number of child spans that were generated while this span was active. If set, allows implementation to detect missing child spans.

span_kind: i32

Optional. Distinguishes between spans generated in a particular context. For example, two spans with the same name may be distinguished using CLIENT (caller) and SERVER (callee) to identify an RPC call.

Implementations

impl Span

pub fn span_kind(&self) -> SpanKind

Returns the enum value of span_kind, or the default if the field is set to an invalid enum value.

pub fn set_span_kind(&mut self, value: SpanKind)

Sets span_kind to the provided enum value.