Struct google_api_proto::google::devtools::resultstore::v2::result_store_download_client::ResultStoreDownloadClient

pub struct ResultStoreDownloadClient<T> { /* private fields */ }

This is the interface used to download information from the ResultStore database.

Clients are encourage to use ExportInvocation for most traffic.

Most APIs require setting a response FieldMask via the ‘fields’ URL query parameter or the X-Goog-FieldMask HTTP/gRPC header.

Implementations

impl<T> ResultStoreDownloadClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> ResultStoreDownloadClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn export_invocation( &mut self, request: impl IntoRequest<ExportInvocationRequest>, ) -> Result<Response<ExportInvocationResponse>, Status>

Exports the invocation with the given name and its child resources.

The order in which resources are returned is defined as follows, invocation; download_metadata; configurations; targets interleaving configured_targets and actions; file_sets.

Invocation

DownloadMetadata

Configurations

Targets └─ ConfiguredTargets └─Actions

FileSets

All child resources will be returned before the next parent resource is returned. For example, all actions under a configured_target will be returned before the next configured_target is returned. The order in which results within a given resource type are returned is undefined, but stable.

An error will be reported in the following cases:

If the invocation is not found.

If the given invocation name is badly formatted.

If no field mask was given.

pub async fn get_invocation( &mut self, request: impl IntoRequest<GetInvocationRequest>, ) -> Result<Response<Invocation>, Status>

Retrieves the invocation with the given name.

An error will be reported in the following cases:

If the invocation is not found.

If the given invocation name is badly formatted.

If no field mask was given.

pub async fn search_invocations( &mut self, request: impl IntoRequest<SearchInvocationsRequest>, ) -> Result<Response<SearchInvocationsResponse>, Status>

Searches for invocations matching the given query parameters. Results will be ordered by timing.start_time with most recent first, but total ordering of results is not guaranteed when difference in timestamps is very small. Results may be stale. Results may be omitted.

An error will be reported in the following cases:

If a query string is not provided

If no field mask was given.

pub async fn get_invocation_download_metadata( &mut self, request: impl IntoRequest<GetInvocationDownloadMetadataRequest>, ) -> Result<Response<DownloadMetadata>, Status>

Retrieves the metadata for an invocation with the given name.

An error will be reported in the following cases:

If the invocation is not found.

If the given invocation name is badly formatted.

pub async fn get_configuration( &mut self, request: impl IntoRequest<GetConfigurationRequest>, ) -> Result<Response<Configuration>, Status>

Retrieves the configuration with the given name.

An error will be reported in the following cases:

If the configuration or its parent invocation is not found.

If the given configuration name is badly formatted.

If no field mask was given.

pub async fn list_configurations( &mut self, request: impl IntoRequest<ListConfigurationsRequest>, ) -> Result<Response<ListConfigurationsResponse>, Status>

Retrieves all configurations for a parent invocation. This might be limited by user or server, in which case a continuation token is provided. The order in which results are returned is undefined, but stable.

An error will be reported in the following cases:

If the parent invocation is not found.

If the given parent invocation name is badly formatted.

If no field mask was given.

pub async fn get_target( &mut self, request: impl IntoRequest<GetTargetRequest>, ) -> Result<Response<Target>, Status>

Retrieves the target with the given name.

An error will be reported in the following cases:

If the target or its parent invocation is not found.

If the given target name is badly formatted.

If no field mask was given.

pub async fn list_targets( &mut self, request: impl IntoRequest<ListTargetsRequest>, ) -> Result<Response<ListTargetsResponse>, Status>

Retrieves all targets for a parent invocation. This might be limited by user or server, in which case a continuation token is provided. The order in which results are returned is undefined, but stable.

An error will be reported in the following cases:

If the parent is not found.

If the given parent name is badly formatted.

If no field mask was given.

pub async fn get_configured_target( &mut self, request: impl IntoRequest<GetConfiguredTargetRequest>, ) -> Result<Response<ConfiguredTarget>, Status>

Retrieves the configured target with the given name.

An error will be reported in the following cases:

If the configured target is not found.

If the given name is badly formatted.

If no field mask was given.

pub async fn list_configured_targets( &mut self, request: impl IntoRequest<ListConfiguredTargetsRequest>, ) -> Result<Response<ListConfiguredTargetsResponse>, Status>

Retrieves all configured targets for a parent invocation/target. This might be limited by user or server, in which case a continuation token is provided. Supports ‘-’ for targetId meaning all targets. The order in which results are returned is undefined, but stable and consistent with ListTargets and ListConfigurations.

An error will be reported in the following cases:

If the parent is not found.

If the given parent name is badly formatted.

If no field mask was given.

pub async fn search_configured_targets( &mut self, request: impl IntoRequest<SearchConfiguredTargetsRequest>, ) -> Result<Response<SearchConfiguredTargetsResponse>, Status>

Searches for ConfiguredTargets matching the given query parameters. Results will be ordered by timing.start_time with most recent first, but total ordering of results is not guaranteed when difference in timestamps is very small. Results may be stale. Results may be omitted.

Field masks are supported for only these fields and their subfields:

configured_targets.name

configured_targets.id

configured_targets.status_attributes

configured_targets.timing

next_page_token

An error will be reported in the following cases:

If a query string is not provided

If no field mask was given.

pub async fn get_action( &mut self, request: impl IntoRequest<GetActionRequest>, ) -> Result<Response<Action>, Status>

Retrieves the action with the given name.

An error will be reported in the following cases:

If the action is not found.

If the given name is badly formatted.

If no field mask was given.

pub async fn list_actions( &mut self, request: impl IntoRequest<ListActionsRequest>, ) -> Result<Response<ListActionsResponse>, Status>

Retrieves all actions for a parent invocation/target/configuration. This might be limited by user or server, in which case a continuation token is provided. Supports ‘-’ for configurationId to mean all actions for all configurations for a target, or ‘-’ for targetId and configurationId to mean all actions for all configurations and all targets. Does not support targetId ‘-’ with a specified configuration. The order in which results are returned is undefined, but stable and consistent with ListConfiguredTargets.

An error will be reported in the following cases:

If the parent is not found.

If the given parent name is badly formatted.

If no field mask was given.

pub async fn batch_list_actions( &mut self, request: impl IntoRequest<BatchListActionsRequest>, ) -> Result<Response<BatchListActionsResponse>, Status>

Retrieves a list of actions for a parent invocation or multiple parents target/configuration. This might be limited by user or server, in which case a continuation token is provided. The order in which results are returned is undefined, but stable and consistent with ListConfiguredTargets.

An error will be reported in the following cases:

If the given parent name is badly formatted.

If no field mask was given.

pub async fn get_file_set( &mut self, request: impl IntoRequest<GetFileSetRequest>, ) -> Result<Response<FileSet>, Status>

Retrieves the file set with the given name.

An error will be reported in the following cases:

If the file set or its parent invocation is not found.

If the given file set name is badly formatted.

If no field mask was given.

pub async fn list_file_sets( &mut self, request: impl IntoRequest<ListFileSetsRequest>, ) -> Result<Response<ListFileSetsResponse>, Status>

Retrieves all file sets for a parent invocation. This might be limited by user or server, in which case a continuation token is provided. The order in which results are returned is undefined, but stable.

An error will be reported in the following cases:

If the parent invocation is not found.

If the given parent invocation name is badly formatted.

If no field mask was given.

pub async fn traverse_file_sets( &mut self, request: impl IntoRequest<TraverseFileSetsRequest>, ) -> Result<Response<TraverseFileSetsResponse>, Status>

Returns the transitive closure of FileSets. This might be limited by user or server, in which case a continuation token is provided. The order in which results are returned is undefined, and unstable.

An error will be reported in the following cases:

If page_token is too large to continue the calculation.

If the resource is not found.

If the given resource name is badly formatted.

If no field mask was given.