Struct google_api_proto::google::devtools::resultstore::v2::Action

pub struct Action {
    pub name: String,
    pub id: Option<Id>,
    pub status_attributes: Option<StatusAttributes>,
    pub timing: Option<Timing>,
    pub action_attributes: Option<ActionAttributes>,
    pub action_dependencies: Vec<Dependency>,
    pub properties: Vec<Property>,
    pub files: Vec<File>,
    pub file_sets: Vec<String>,
    pub coverage: Option<ActionCoverage>,
    pub file_processing_errors: Vec<FileProcessingErrors>,
    pub action_type: Option<ActionType>,
}

An action that happened as part of a configured target. This action could be a build, a test, or another type of action, as specified in action_type oneof.

Each parent ConfiguredTarget resource should have at least one Action as its child resource before the invocation is finalized. For a simple build, at least one build action should be created to represent the build result, and at least one test action should be created to represent the test result, if any.

Fields
name: String

The resource name. Its format must be: invocations/${INVOCATION_ID}/targets/${url_encode(TARGET_ID)}/configuredTargets/url_encode(${CONFIG_ID})/actions/${url_encode(ACTION_ID)}

See CreateActionRequest proto for more information.

id: Option<Id>

The resource ID components that identify the Action. They must match the resource name after proper encoding.

status_attributes: Option<StatusAttributes>

The status of the action.

timing: Option<Timing>

The timing of the whole action. For TestActions, the start time may be before the test actually started, and the duration may last until after the test actually finished.

action_attributes: Option<ActionAttributes>

General attributes of the action.

action_dependencies: Vec<Dependency>

A list of resources that this action depended upon. May be used to provide the cause of a build failure in the case of a failed build action.

properties: Vec<Property>

Arbitrary name-value pairs. This is implemented as a multi-map. Multiple properties are allowed with the same key. Properties will be returned in lexicographical order by key.

files: Vec<File>

A list of file references for action level files. The file IDs must be unique within this list. Duplicate file IDs will result in an error. Files will be returned in lexicographical order by ID.

Files with the following reserved file IDs cause specific post-processing or have special handling. These files must be immediately available to ResultStore for processing when the reference is uploaded.

For build actions: stdout: The stdout of the action stderr: The stderr of the action baseline.lcov: Baseline coverage file to be parsed by the server. This uses a stripped down implementation of the LCOV standard. http://ltp.sourceforge.net/coverage/lcov/geninfo.1.php

For test actions: test.xml: The test suite / test case data in XML format. test.log: The combined stdout and stderr of the test process. test.lcov: Coverage file to be parsed by the server. This uses a stripped down implementation of the LCOV standard. http://ltp.sourceforge.net/coverage/lcov/geninfo.1.php

file_sets: Vec<String>

List of names of file sets that are referenced from this Action. Each name must point to a file set under the same Invocation. The name format must be: invocations/${INVOCATION_ID}/fileSets/${FILE_SET_ID}

coverage: Option<ActionCoverage>

Coverage data was collected while running the build or test action. This usually includes line coverage, and may also include branch coverage. For test actions, this is usually only for the source files which were actually executed by that particular action. For build actions, this is the baseline coverage, which captures the instrumented files and lines, without any lines being executed. This ensures files that are never covered at all are included.

file_processing_errors: Vec<FileProcessingErrors>

ResultStore will read and parse Files with reserved IDs listed above. Read and parse errors for all these Files are reported here. This is implemented as a map, with one FileProcessingErrors for each file. Typically produced when parsing Files, but may also be provided directly by clients.

action_type: Option<ActionType>

The type of the action. The type of an action may not change over the lifetime of the invocation. If one of these fields is to be set, it must be set in the CreateAction method. It may be set to an empty message that is populated in later methods or post-processing. A generic “untyped” action can be created by not setting any of these fields. An untyped action will be untyped for the lifetime of the invocation.