Module google_api_proto::google::firestore::admin::v1

Modules

backup

Nested message and enum types in Backup.

backup_schedule

Nested message and enum types in BackupSchedule.

database

Nested message and enum types in Database.

field

Nested message and enum types in Field.

field_operation_metadata

Nested message and enum types in FieldOperationMetadata.

firestore_admin_client

Generated client implementations.

index

Nested message and enum types in Index.

Structs

Backup

A Backup of a Cloud Firestore Database.

BackupSchedule

A backup schedule for a Cloud Firestore Database.

BulkDeleteDocumentsMetadata

Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.BulkDeleteDocuments][google.firestore.admin.v1.FirestoreAdmin.BulkDeleteDocuments].

BulkDeleteDocumentsRequest

The request for [FirestoreAdmin.BulkDeleteDocuments][google.firestore.admin.v1.FirestoreAdmin.BulkDeleteDocuments].

BulkDeleteDocumentsResponse

The response for [FirestoreAdmin.BulkDeleteDocuments][google.firestore.admin.v1.FirestoreAdmin.BulkDeleteDocuments].

CreateBackupScheduleRequest

The request for [FirestoreAdmin.CreateBackupSchedule][google.firestore.admin.v1.FirestoreAdmin.CreateBackupSchedule].

CreateDatabaseMetadata

Metadata related to the create database operation.

CreateDatabaseRequest

The request for [FirestoreAdmin.CreateDatabase][google.firestore.admin.v1.FirestoreAdmin.CreateDatabase].

CreateIndexRequest

The request for [FirestoreAdmin.CreateIndex][google.firestore.admin.v1.FirestoreAdmin.CreateIndex].

DailyRecurrence

Represents a recurring schedule that runs at a specific time every day.

Database

A Cloud Firestore Database.

DeleteBackupRequest

The request for [FirestoreAdmin.DeleteBackup][google.firestore.admin.v1.FirestoreAdmin.DeleteBackup].

DeleteBackupScheduleRequest

The request for [FirestoreAdmin.DeleteBackupSchedules][].

DeleteDatabaseMetadata

Metadata related to the delete database operation.

DeleteDatabaseRequest

The request for [FirestoreAdmin.DeleteDatabase][google.firestore.admin.v1.FirestoreAdmin.DeleteDatabase].

DeleteIndexRequest

The request for [FirestoreAdmin.DeleteIndex][google.firestore.admin.v1.FirestoreAdmin.DeleteIndex].

ExportDocumentsMetadata

Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.ExportDocuments][google.firestore.admin.v1.FirestoreAdmin.ExportDocuments].

ExportDocumentsRequest

The request for [FirestoreAdmin.ExportDocuments][google.firestore.admin.v1.FirestoreAdmin.ExportDocuments].

ExportDocumentsResponse

Returned in the [google.longrunning.Operation][google.longrunning.Operation] response field.

Field

Represents a single field in the database.

FieldOperationMetadata

Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.UpdateField][google.firestore.admin.v1.FirestoreAdmin.UpdateField].

GetBackupRequest

The request for [FirestoreAdmin.GetBackup][google.firestore.admin.v1.FirestoreAdmin.GetBackup].

GetBackupScheduleRequest

The request for [FirestoreAdmin.GetBackupSchedule][google.firestore.admin.v1.FirestoreAdmin.GetBackupSchedule].

GetDatabaseRequest

The request for [FirestoreAdmin.GetDatabase][google.firestore.admin.v1.FirestoreAdmin.GetDatabase].

GetFieldRequest

The request for [FirestoreAdmin.GetField][google.firestore.admin.v1.FirestoreAdmin.GetField].

GetIndexRequest

The request for [FirestoreAdmin.GetIndex][google.firestore.admin.v1.FirestoreAdmin.GetIndex].

ImportDocumentsMetadata

Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.ImportDocuments][google.firestore.admin.v1.FirestoreAdmin.ImportDocuments].

ImportDocumentsRequest

The request for [FirestoreAdmin.ImportDocuments][google.firestore.admin.v1.FirestoreAdmin.ImportDocuments].

Index

Cloud Firestore indexes enable simple and complex queries against documents in a database.

IndexOperationMetadata

Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.CreateIndex][google.firestore.admin.v1.FirestoreAdmin.CreateIndex].

ListBackupSchedulesRequest

The request for [FirestoreAdmin.ListBackupSchedules][google.firestore.admin.v1.FirestoreAdmin.ListBackupSchedules].

ListBackupSchedulesResponse

The response for [FirestoreAdmin.ListBackupSchedules][google.firestore.admin.v1.FirestoreAdmin.ListBackupSchedules].

ListBackupsRequest

The request for [FirestoreAdmin.ListBackups][google.firestore.admin.v1.FirestoreAdmin.ListBackups].

ListBackupsResponse

The response for [FirestoreAdmin.ListBackups][google.firestore.admin.v1.FirestoreAdmin.ListBackups].

ListDatabasesRequest

A request to list the Firestore Databases in all locations for a project.

ListDatabasesResponse

The list of databases for a project.

ListFieldsRequest

The request for [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields].

ListFieldsResponse

The response for [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields].

ListIndexesRequest

The request for [FirestoreAdmin.ListIndexes][google.firestore.admin.v1.FirestoreAdmin.ListIndexes].

ListIndexesResponse

The response for [FirestoreAdmin.ListIndexes][google.firestore.admin.v1.FirestoreAdmin.ListIndexes].

LocationMetadata

The metadata message for [google.cloud.location.Location.metadata][google.cloud.location.Location.metadata].

Progress

Describes the progress of the operation. Unit of work is generic and must be interpreted based on where [Progress][google.firestore.admin.v1.Progress] is used.

RestoreDatabaseMetadata

Metadata for the [long-running operation][google.longrunning.Operation] from the [RestoreDatabase][google.firestore.admin.v1.RestoreDatabase] request.

RestoreDatabaseRequest

The request message for [FirestoreAdmin.RestoreDatabase][google.firestore.admin.v1.RestoreDatabase].

UpdateBackupScheduleRequest

The request for [FirestoreAdmin.UpdateBackupSchedule][google.firestore.admin.v1.FirestoreAdmin.UpdateBackupSchedule].

UpdateDatabaseMetadata

Metadata related to the update database operation.

UpdateDatabaseRequest

The request for [FirestoreAdmin.UpdateDatabase][google.firestore.admin.v1.FirestoreAdmin.UpdateDatabase].

UpdateFieldRequest

The request for [FirestoreAdmin.UpdateField][google.firestore.admin.v1.FirestoreAdmin.UpdateField].

WeeklyRecurrence

Represents a recurring schedule that runs on a specified day of the week.

Enums

OperationState

Describes the state of the operation.