Struct google_api_proto::google::firestore::admin::v1::firestore_admin_client::FirestoreAdminClient

pub struct FirestoreAdminClient<T> { /* private fields */ }

The Cloud Firestore Admin API.

This API provides several administrative services for Cloud Firestore.

Project, Database, Namespace, Collection, Collection Group, and Document are used as defined in the Google Cloud Firestore API.

Operation: An Operation represents work being performed in the background.

The index service manages Cloud Firestore indexes.

Index creation is performed asynchronously. An Operation resource is created for each such asynchronous operation. The state of the operation (including any errors encountered) may be queried via the Operation resource.

The Operations collection provides a record of actions performed for the specified Project (including any Operations in progress). Operations are not created directly but through calls on other collections or resources.

An Operation that is done may be deleted so that it is no longer listed as part of the Operation collection. Operations are garbage collected after 30 days. By default, ListOperations will only return in progress and failed operations. To list completed operation, issue a ListOperations request with the filter done: true.

Operations are created by service FirestoreAdmin, but are accessed via service google.longrunning.Operations.

Implementations

impl<T> FirestoreAdminClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> FirestoreAdminClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_index( &mut self, request: impl IntoRequest<CreateIndexRequest>, ) -> Result<Response<Operation>, Status>

Creates a composite index. This returns a [google.longrunning.Operation][google.longrunning.Operation] which may be used to track the status of the creation. The metadata for the operation will be the type [IndexOperationMetadata][google.firestore.admin.v1.IndexOperationMetadata].

pub async fn list_indexes( &mut self, request: impl IntoRequest<ListIndexesRequest>, ) -> Result<Response<ListIndexesResponse>, Status>

Lists composite indexes.

pub async fn get_index( &mut self, request: impl IntoRequest<GetIndexRequest>, ) -> Result<Response<Index>, Status>

Gets a composite index.

pub async fn delete_index( &mut self, request: impl IntoRequest<DeleteIndexRequest>, ) -> Result<Response<()>, Status>

Deletes a composite index.

pub async fn get_field( &mut self, request: impl IntoRequest<GetFieldRequest>, ) -> Result<Response<Field>, Status>

Gets the metadata and configuration for a Field.

pub async fn update_field( &mut self, request: impl IntoRequest<UpdateFieldRequest>, ) -> Result<Response<Operation>, Status>

Updates a field configuration. Currently, field updates apply only to single field index configuration. However, calls to [FirestoreAdmin.UpdateField][google.firestore.admin.v1.FirestoreAdmin.UpdateField] should provide a field mask to avoid changing any configuration that the caller isn’t aware of. The field mask should be specified as: { paths: "index_config" }.

This call returns a [google.longrunning.Operation][google.longrunning.Operation] which may be used to track the status of the field update. The metadata for the operation will be the type [FieldOperationMetadata][google.firestore.admin.v1.FieldOperationMetadata].

To configure the default field settings for the database, use the special Field with resource name: projects/{project_id}/databases/{database_id}/collectionGroups/__default__/fields/*.

pub async fn list_fields( &mut self, request: impl IntoRequest<ListFieldsRequest>, ) -> Result<Response<ListFieldsResponse>, Status>

Lists the field configuration and metadata for this database.

Currently, [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields] only supports listing fields that have been explicitly overridden. To issue this query, call [FirestoreAdmin.ListFields][google.firestore.admin.v1.FirestoreAdmin.ListFields] with the filter set to indexConfig.usesAncestorConfig:false or ttlConfig:*.

pub async fn export_documents( &mut self, request: impl IntoRequest<ExportDocumentsRequest>, ) -> Result<Response<Operation>, Status>

Exports a copy of all or a subset of documents from Google Cloud Firestore to another storage system, such as Google Cloud Storage. Recent updates to documents may not be reflected in the export. The export occurs in the background and its progress can be monitored and managed via the Operation resource that is created. The output of an export may only be used once the associated operation is done. If an export operation is cancelled before completion it may leave partial data behind in Google Cloud Storage.

For more details on export behavior and output format, refer to: https://cloud.google.com/firestore/docs/manage-data/export-import

pub async fn import_documents( &mut self, request: impl IntoRequest<ImportDocumentsRequest>, ) -> Result<Response<Operation>, Status>

Imports documents into Google Cloud Firestore. Existing documents with the same name are overwritten. The import occurs in the background and its progress can be monitored and managed via the Operation resource that is created. If an ImportDocuments operation is cancelled, it is possible that a subset of the data has already been imported to Cloud Firestore.

pub async fn bulk_delete_documents( &mut self, request: impl IntoRequest<BulkDeleteDocumentsRequest>, ) -> Result<Response<Operation>, Status>

Bulk deletes a subset of documents from Google Cloud Firestore. Documents created or updated after the underlying system starts to process the request will not be deleted. The bulk delete occurs in the background and its progress can be monitored and managed via the Operation resource that is created.

For more details on bulk delete behavior, refer to: https://cloud.google.com/firestore/docs/manage-data/bulk-delete

pub async fn create_database( &mut self, request: impl IntoRequest<CreateDatabaseRequest>, ) -> Result<Response<Operation>, Status>

Create a database.

pub async fn get_database( &mut self, request: impl IntoRequest<GetDatabaseRequest>, ) -> Result<Response<Database>, Status>

Gets information about a database.

pub async fn list_databases( &mut self, request: impl IntoRequest<ListDatabasesRequest>, ) -> Result<Response<ListDatabasesResponse>, Status>

List all the databases in the project.

pub async fn update_database( &mut self, request: impl IntoRequest<UpdateDatabaseRequest>, ) -> Result<Response<Operation>, Status>

Updates a database.

pub async fn delete_database( &mut self, request: impl IntoRequest<DeleteDatabaseRequest>, ) -> Result<Response<Operation>, Status>

Deletes a database.

pub async fn get_backup( &mut self, request: impl IntoRequest<GetBackupRequest>, ) -> Result<Response<Backup>, Status>

Gets information about a backup.

pub async fn list_backups( &mut self, request: impl IntoRequest<ListBackupsRequest>, ) -> Result<Response<ListBackupsResponse>, Status>

Lists all the backups.

pub async fn delete_backup( &mut self, request: impl IntoRequest<DeleteBackupRequest>, ) -> Result<Response<()>, Status>

Deletes a backup.

pub async fn restore_database( &mut self, request: impl IntoRequest<RestoreDatabaseRequest>, ) -> Result<Response<Operation>, Status>

Creates a new database by restoring from an existing backup.

The new database must be in the same cloud region or multi-region location as the existing backup. This behaves similar to [FirestoreAdmin.CreateDatabase][google.firestore.admin.v1.FirestoreAdmin.CreateDatabase] except instead of creating a new empty database, a new database is created with the database type, index configuration, and documents from an existing backup.

The [long-running operation][google.longrunning.Operation] can be used to track the progress of the restore, with the Operation’s [metadata][google.longrunning.Operation.metadata] field type being the [RestoreDatabaseMetadata][google.firestore.admin.v1.RestoreDatabaseMetadata]. The [response][google.longrunning.Operation.response] type is the [Database][google.firestore.admin.v1.Database] if the restore was successful. The new database is not readable or writeable until the LRO has completed.

pub async fn create_backup_schedule( &mut self, request: impl IntoRequest<CreateBackupScheduleRequest>, ) -> Result<Response<BackupSchedule>, Status>

Creates a backup schedule on a database. At most two backup schedules can be configured on a database, one daily backup schedule and one weekly backup schedule.

pub async fn get_backup_schedule( &mut self, request: impl IntoRequest<GetBackupScheduleRequest>, ) -> Result<Response<BackupSchedule>, Status>

Gets information about a backup schedule.

pub async fn list_backup_schedules( &mut self, request: impl IntoRequest<ListBackupSchedulesRequest>, ) -> Result<Response<ListBackupSchedulesResponse>, Status>

List backup schedules.

pub async fn update_backup_schedule( &mut self, request: impl IntoRequest<UpdateBackupScheduleRequest>, ) -> Result<Response<BackupSchedule>, Status>

Updates a backup schedule.

pub async fn delete_backup_schedule( &mut self, request: impl IntoRequest<DeleteBackupScheduleRequest>, ) -> Result<Response<()>, Status>

Deletes a backup schedule.