Struct google_api_proto::google::firestore::admin::v1::Database

pub struct Database {

    pub name: String,
    pub uid: String,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub location_id: String,
    pub type: i32,
    pub concurrency_mode: i32,
    pub version_retention_period: Option<Duration>,
    pub earliest_version_time: Option<Timestamp>,
    pub point_in_time_recovery_enablement: i32,
    pub app_engine_integration_mode: i32,
    pub key_prefix: String,
    pub delete_protection_state: i32,
    pub etag: String,

}

A Cloud Firestore Database.

Fields
name: String

The resource name of the Database. Format: projects/{project}/databases/{database}

uid: String

Output only. The system-generated UUID4 for this Database.

create_time: Option<Timestamp>

Output only. The timestamp at which this database was created. Databases created before 2016 do not populate create_time.

update_time: Option<Timestamp>

Output only. The timestamp at which this database was most recently updated. Note this only includes updates to the database resource and not data contained by the database.

location_id: String

The location of the database. Available locations are listed at https://cloud.google.com/firestore/docs/locations.

type: i32

The type of the database. See https://cloud.google.com/datastore/docs/firestore-or-datastore for information about how to choose.

concurrency_mode: i32

The concurrency control mode to use for this database.

version_retention_period: Option<Duration>

Output only. The period during which past versions of data are retained in the database.

Any [read][google.firestore.v1.GetDocumentRequest.read_time] or [query][google.firestore.v1.ListDocumentsRequest.read_time] can specify a read_time within this window, and will read the state of the database at that time.

If the PITR feature is enabled, the retention period is 7 days. Otherwise, the retention period is 1 hour.

earliest_version_time: Option<Timestamp>

Output only. The earliest timestamp at which older versions of the data can be read from the database. See [version_retention_period] above; this field is populated with now - version_retention_period.

This value is continuously updated, and becomes stale the moment it is queried. If you are using this value to recover data, make sure to account for the time from the moment when the value is queried to the moment when you initiate the recovery.

point_in_time_recovery_enablement: i32

Whether to enable the PITR feature on this database.

app_engine_integration_mode: i32

The App Engine integration mode to use for this database.

key_prefix: String

Output only. The key_prefix for this database. This key_prefix is used, in combination with the project id (“~”) to construct the application id that is returned from the Cloud Datastore APIs in Google App Engine first generation runtimes.

This value may be empty in which case the appid to use for URL-encoded keys is the project_id (eg: foo instead of v~foo).

delete_protection_state: i32

State of delete protection for the database.

etag: String

This checksum is computed by the server based on the value of other fields, and may be sent on update and delete requests to ensure the client has an up-to-date value before proceeding.

Implementations

impl Database

pub fn type(&self) -> DatabaseType

Returns the enum value of type, or the default if the field is set to an invalid enum value.

pub fn set_type(&mut self, value: DatabaseType)

Sets type to the provided enum value.

pub fn concurrency_mode(&self) -> ConcurrencyMode

Returns the enum value of concurrency_mode, or the default if the field is set to an invalid enum value.

pub fn set_concurrency_mode(&mut self, value: ConcurrencyMode)

Sets concurrency_mode to the provided enum value.

pub fn app_engine_integration_mode(&self) -> AppEngineIntegrationMode

Returns the enum value of app_engine_integration_mode, or the default if the field is set to an invalid enum value.

pub fn set_app_engine_integration_mode( &mut self, value: AppEngineIntegrationMode, )

Sets app_engine_integration_mode to the provided enum value.

pub fn point_in_time_recovery_enablement(&self) -> PointInTimeRecoveryEnablement

Returns the enum value of point_in_time_recovery_enablement, or the default if the field is set to an invalid enum value.

pub fn set_point_in_time_recovery_enablement( &mut self, value: PointInTimeRecoveryEnablement, )

Sets point_in_time_recovery_enablement to the provided enum value.

pub fn delete_protection_state(&self) -> DeleteProtectionState

Returns the enum value of delete_protection_state, or the default if the field is set to an invalid enum value.

pub fn set_delete_protection_state(&mut self, value: DeleteProtectionState)

Sets delete_protection_state to the provided enum value.