Module google_api_proto::google::firestore::bundle

Modules

bundle_element

Nested message and enum types in BundleElement.

bundled_query

Nested message and enum types in BundledQuery.

Structs

BundleElement

A Firestore bundle is a length-prefixed stream of JSON representations of BundleElement. Only one BundleMetadata is expected, and it should be the first element. The named queries follow after metadata. Every document_metadata is immediately followed by a document.

BundleMetadata

Metadata describing the bundle file/stream.

BundledDocumentMetadata

Metadata describing a Firestore document saved in the bundle.

BundledQuery

Encodes a query saved in the bundle.

NamedQuery

A Query associated with a name, created as part of the bundle file, and can be read by client SDKs once the bundle containing them is loaded.