Struct google_api_proto::google::genomics::v1::read_service_v1_client::ReadServiceV1Client

pub struct ReadServiceV1Client<T> { /* private fields */ }

The Readstore. A data store for DNA sequencing Reads.

Implementations

impl<T> ReadServiceV1Client<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> ReadServiceV1Client<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn import_read_group_sets( &mut self, request: impl IntoRequest<ImportReadGroupSetsRequest>, ) -> Result<Response<Operation>, Status>

Creates read group sets by asynchronously importing the provided information.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

The caller must have WRITE permissions to the dataset.

Notes on BAM import

Tags will be converted to strings - tag types are not preserved

Comments (@CO) in the input file header will not be preserved

Original header order of references (@SQ) will not be preserved

Any reverse stranded unmapped reads will be reverse complemented, and their qualities (also the “BQ” and “OQ” tags, if any) will be reversed

Unmapped reads will be stripped of positional information (reference name and position)

pub async fn export_read_group_set( &mut self, request: impl IntoRequest<ExportReadGroupSetRequest>, ) -> Result<Response<Operation>, Status>

Exports a read group set to a BAM file in Google Cloud Storage.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Note that currently there may be some differences between exported BAM files and the original BAM file at the time of import. See [ImportReadGroupSets][google.genomics.v1.ReadServiceV1.ImportReadGroupSets] for caveats.

pub async fn search_read_group_sets( &mut self, request: impl IntoRequest<SearchReadGroupSetsRequest>, ) -> Result<Response<SearchReadGroupSetsResponse>, Status>

Searches for read group sets matching the criteria.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Implements GlobalAllianceApi.searchReadGroupSets.

pub async fn update_read_group_set( &mut self, request: impl IntoRequest<UpdateReadGroupSetRequest>, ) -> Result<Response<ReadGroupSet>, Status>

Updates a read group set.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

This method supports patch semantics.

pub async fn delete_read_group_set( &mut self, request: impl IntoRequest<DeleteReadGroupSetRequest>, ) -> Result<Response<()>, Status>

Deletes a read group set.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

pub async fn get_read_group_set( &mut self, request: impl IntoRequest<GetReadGroupSetRequest>, ) -> Result<Response<ReadGroupSet>, Status>

Gets a read group set by ID.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

pub async fn list_coverage_buckets( &mut self, request: impl IntoRequest<ListCoverageBucketsRequest>, ) -> Result<Response<ListCoverageBucketsResponse>, Status>

Lists fixed width coverage buckets for a read group set, each of which correspond to a range of a reference sequence. Each bucket summarizes coverage information across its corresponding genomic range.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Coverage is defined as the number of reads which are aligned to a given base in the reference sequence. Coverage buckets are available at several precomputed bucket widths, enabling retrieval of various coverage ‘zoom levels’. The caller must have READ permissions for the target read group set.

pub async fn search_reads( &mut self, request: impl IntoRequest<SearchReadsRequest>, ) -> Result<Response<SearchReadsResponse>, Status>

Gets a list of reads for one or more read group sets.

For the definitions of read group sets and other genomics resources, see Fundamentals of Google Genomics

Reads search operates over a genomic coordinate space of reference sequence & position defined over the reference sequences to which the requested read group sets are aligned.

If a target positional range is specified, search returns all reads whose alignment to the reference genome overlap the range. A query which specifies only read group set IDs yields all reads in those read group sets, including unmapped reads.

All reads returned (including reads on subsequent pages) are ordered by genomic coordinate (by reference sequence, then position). Reads with equivalent genomic coordinates are returned in an unspecified order. This order is consistent, such that two queries for the same content (regardless of page size) yield reads in the same order across their respective streams of paginated responses.

Implements GlobalAllianceApi.searchReads.