Struct google_api_proto::google::genomics::v1::Variant

pub struct Variant {

    pub variant_set_id: String,
    pub id: String,
    pub names: Vec<String>,
    pub created: i64,
    pub reference_name: String,
    pub start: i64,
    pub end: i64,
    pub reference_bases: String,
    pub alternate_bases: Vec<String>,
    pub quality: f64,
    pub filter: Vec<String>,
    pub info: BTreeMap<String, ListValue>,
    pub calls: Vec<VariantCall>,

}

A variant represents a change in DNA sequence relative to a reference sequence. For example, a variant could represent a SNP or an insertion. Variants belong to a variant set.

For more genomics resource definitions, see Fundamentals of Google Genomics

Each of the calls on a variant represent a determination of genotype with respect to that variant. For example, a call might assign probability of 0.32 to the occurrence of a SNP named rs1234 in a sample named NA12345. A call belongs to a call set, which contains related calls typically from one sample.

Fields
variant_set_id: String

The ID of the variant set this variant belongs to.

id: String

The server-generated variant ID, unique across all variants.

names: Vec<String>

Names for the variant, for example a RefSNP ID.

created: i64

The date this variant was created, in milliseconds from the epoch.

reference_name: String

The reference on which this variant occurs. (such as chr20 or X)

start: i64

The position at which this variant occurs (0-based). This corresponds to the first base of the string of reference bases.

end: i64

The end position (0-based) of this variant. This corresponds to the first base after the last base in the reference allele. So, the length of the reference allele is (end - start). This is useful for variants that don’t explicitly give alternate bases, for example large deletions.

reference_bases: String

The reference bases for this variant. They start at the given position.

alternate_bases: Vec<String>

The bases that appear instead of the reference bases.

quality: f64

A measure of how likely this variant is to be real. A higher value is better.

filter: Vec<String>

A list of filters (normally quality filters) this variant has failed. PASS indicates this variant has passed all filters.

info: BTreeMap<String, ListValue>

A map of additional variant information. This must be of the form map<string, string[]> (string key mapping to a list of string values).

calls: Vec<VariantCall>

The variant calls for this particular variant. Each one represents the determination of genotype with respect to this variant.