Module google_api_proto::google::genomics::v1alpha2

Modules

controller_config

Nested message and enum types in ControllerConfig.

pipeline

Nested message and enum types in Pipeline.

pipeline_parameter

Nested message and enum types in PipelineParameter.

pipeline_resources

Nested message and enum types in PipelineResources.

pipelines_v1_alpha2_client

Generated client implementations.

run_pipeline_request

Nested message and enum types in RunPipelineRequest.

Structs

ComputeEngine

Describes a Compute Engine resource that is being managed by a running [pipeline][google.genomics.v1alpha2.Pipeline].

ControllerConfig

Stores the information that the controller will fetch from the server in order to run. Should only be used by VMs created by the Pipelines Service and not by end users.

CreatePipelineRequest

The request to create a pipeline. The pipeline field here should not have pipelineId populated, as that will be populated by the server.

DeletePipelineRequest

The request to delete a saved pipeline by ID.

DockerExecutor

The Docker execuctor specification.

GetControllerConfigRequest

Request to get controller configuation. Should only be used by VMs created by the Pipelines Service and not by end users.

GetPipelineRequest

A request to get a saved pipeline by id.

ListPipelinesRequest

A request to list pipelines in a given project. Pipelines can be filtered by name using namePrefix: all pipelines with names that begin with namePrefix will be returned. Uses standard pagination: pageSize indicates how many pipelines to return, and pageToken comes from a previous ListPipelinesResponse to indicate offset.

ListPipelinesResponse

The response of ListPipelines. Contains at most pageSize pipelines. If it contains pageSize pipelines, and more pipelines exist, then nextPageToken will be populated and should be used as the pageToken argument to a subsequent ListPipelines request.

LoggingOptions

The logging options for the pipeline run.

Pipeline

The pipeline object. Represents a transformation from a set of input parameters to a set of output parameters. The transformation is defined as a docker image and command to run within that image. Each pipeline is run on a Google Compute Engine VM. A pipeline can be created with the create method and then later run with the run method, or a pipeline can be defined and run all at once with the run method.

PipelineParameter

Parameters facilitate setting and delivering data into the pipeline’s execution environment. They are defined at create time, with optional defaults, and can be overridden at run time.

PipelineResources

The system resources for the pipeline run.

RunPipelineArgs

The pipeline run arguments.

RunPipelineRequest

The request to run a pipeline. If pipelineId is specified, it refers to a saved pipeline created with CreatePipeline and set as the pipelineId of the returned Pipeline object. If ephemeralPipeline is specified, that pipeline is run once with the given args and not saved. It is an error to specify both pipelineId and ephemeralPipeline. pipelineArgs must be specified.

RuntimeMetadata

Runtime metadata that will be populated in the [runtimeMetadata][google.genomics.v1.OperationMetadata.runtime_metadata] field of the Operation associated with a RunPipeline execution.

ServiceAccount

A Google Cloud Service Account.

SetOperationStatusRequest

Request to set operation status. Should only be used by VMs created by the Pipelines Service and not by end users.

TimestampEvent

Stores the list of events and times they occured for major events in job execution.