Module google_api_proto::google::logging::v2

Modules

bucket_metadata

Nested message and enum types in BucketMetadata.

config_service_v2_client

Generated client implementations.

link_metadata

Nested message and enum types in LinkMetadata.

log_entry

Nested message and enum types in LogEntry.

log_metric

Nested message and enum types in LogMetric.

log_sink

Nested message and enum types in LogSink.

logging_service_v2_client

Generated client implementations.

metrics_service_v2_client

Generated client implementations.

tail_log_entries_response

Nested message and enum types in TailLogEntriesResponse.

Structs

BigQueryDataset

Describes a BigQuery dataset that was created by a link.

BigQueryOptions

Options that change functionality of a sink exporting data to BigQuery.

BucketMetadata

Metadata for LongRunningUpdateBucket Operations.

CmekSettings

Describes the customer-managed encryption key (CMEK) settings associated with a project, folder, organization, billing account, or flexible resource.

CopyLogEntriesMetadata

Metadata for CopyLogEntries long running operations.

CopyLogEntriesRequest

The parameters to CopyLogEntries.

CopyLogEntriesResponse

Response type for CopyLogEntries long running operations.

CreateBucketRequest

The parameters to CreateBucket.

CreateExclusionRequest

The parameters to CreateExclusion.

CreateLinkRequest

The parameters to CreateLink.

CreateLogMetricRequest

The parameters to CreateLogMetric.

CreateSinkRequest

The parameters to CreateSink.

CreateViewRequest

The parameters to CreateView.

DeleteBucketRequest

The parameters to DeleteBucket.

DeleteExclusionRequest

The parameters to DeleteExclusion.

DeleteLinkRequest

The parameters to DeleteLink.

DeleteLogMetricRequest

The parameters to DeleteLogMetric.

DeleteLogRequest

The parameters to DeleteLog.

DeleteSinkRequest

The parameters to DeleteSink.

DeleteViewRequest

The parameters to DeleteView.

GetBucketRequest

The parameters to GetBucket.

GetCmekSettingsRequest

The parameters to [GetCmekSettings][google.logging.v2.ConfigServiceV2.GetCmekSettings].

GetExclusionRequest

The parameters to GetExclusion.

GetLinkRequest

The parameters to GetLink.

GetLogMetricRequest

The parameters to GetLogMetric.

GetSettingsRequest

The parameters to [GetSettings][google.logging.v2.ConfigServiceV2.GetSettings].

GetSinkRequest

The parameters to GetSink.

GetViewRequest

The parameters to GetView.

IndexConfig

Configuration for an indexed field.

Link

Describes a link connected to an analytics enabled bucket.

LinkMetadata

Metadata for long running Link operations.

ListBucketsRequest

The parameters to ListBuckets.

ListBucketsResponse

The response from ListBuckets.

ListExclusionsRequest

The parameters to ListExclusions.

ListExclusionsResponse

Result returned from ListExclusions.

ListLinksRequest

The parameters to ListLinks.

ListLinksResponse

The response from ListLinks.

ListLogEntriesRequest

The parameters to ListLogEntries.

ListLogEntriesResponse

Result returned from ListLogEntries.

ListLogMetricsRequest

The parameters to ListLogMetrics.

ListLogMetricsResponse

Result returned from ListLogMetrics.

ListLogsRequest

The parameters to ListLogs.

ListLogsResponse

Result returned from ListLogs.

ListMonitoredResourceDescriptorsRequest

The parameters to ListMonitoredResourceDescriptors

ListMonitoredResourceDescriptorsResponse

Result returned from ListMonitoredResourceDescriptors.

ListSinksRequest

The parameters to ListSinks.

ListSinksResponse

Result returned from ListSinks.

ListViewsRequest

The parameters to ListViews.

ListViewsResponse

The response from ListViews.

LocationMetadata

Cloud Logging specific location metadata.

LogBucket

Describes a repository in which log entries are stored.

LogEntry

An individual entry in a log.

LogEntryOperation

Additional information about a potentially long-running operation with which a log entry is associated.

LogEntrySourceLocation

Additional information about the source code location that produced the log entry.

LogExclusion

Specifies a set of log entries that are filtered out by a sink. If your Google Cloud resource receives a large volume of log entries, you can use exclusions to reduce your chargeable logs. Note that exclusions on organization-level and folder-level sinks don’t apply to child resources. Note also that you cannot modify the _Required sink or exclude logs from it.

LogMetric

Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter in a given time interval.

LogSink

Describes a sink used to export log entries to one of the following destinations in any project: a Cloud Storage bucket, a BigQuery dataset, a Pub/Sub topic or a Cloud Logging log bucket. A logs filter controls which log entries are exported. The sink must be created within a project, organization, billing account, or folder.

LogSplit

Additional information used to correlate multiple log entries. Used when a single LogEntry would exceed the Google Cloud Logging size limit and is split across multiple log entries.

LogView

Describes a view over log entries in a bucket.

Settings

Describes the settings associated with a project, folder, organization, billing account, or flexible resource.

TailLogEntriesRequest

The parameters to TailLogEntries.

TailLogEntriesResponse

Result returned from TailLogEntries.

UndeleteBucketRequest

The parameters to UndeleteBucket.

UpdateBucketRequest

The parameters to UpdateBucket.

UpdateCmekSettingsRequest

The parameters to [UpdateCmekSettings][google.logging.v2.ConfigServiceV2.UpdateCmekSettings].

UpdateExclusionRequest

The parameters to UpdateExclusion.

UpdateLogMetricRequest

The parameters to UpdateLogMetric.

UpdateSettingsRequest

The parameters to [UpdateSettings][google.logging.v2.ConfigServiceV2.UpdateSettings].

UpdateSinkRequest

The parameters to UpdateSink.

UpdateViewRequest

The parameters to UpdateView.

WriteLogEntriesPartialErrors

Error details for WriteLogEntries with partial success.

WriteLogEntriesRequest

The parameters to WriteLogEntries.

WriteLogEntriesResponse

Result returned from WriteLogEntries.

Enums

IndexType

IndexType is used for custom indexing. It describes the type of an indexed field.

LifecycleState

LogBucket lifecycle states.

OperationState

List of different operation states. High level state of the operation. This is used to report the job’s current state to the user. Once a long running operation is created, the current state of the operation can be queried even before the operation is finished and the final result is available.