Struct google_api_proto::google::logging::v2::LogMetric

pub struct LogMetric {
    pub name: String,
    pub description: String,
    pub filter: String,
    pub bucket_name: String,
    pub disabled: bool,
    pub metric_descriptor: Option<MetricDescriptor>,
    pub value_extractor: String,
    pub label_extractors: BTreeMap<String, String>,
    pub bucket_options: Option<BucketOptions>,
    pub create_time: Option<Timestamp>,
    pub update_time: Option<Timestamp>,
    pub version: i32,
}

Describes a logs-based metric. The value of the metric is the number of log entries that match a logs filter in a given time interval.

Logs-based metrics can also be used to extract values from logs and create a distribution of the values. The distribution records the statistics of the extracted values along with an optional histogram of the values as specified by the bucket options.

Fields
name: String

Required. The client-assigned metric identifier. Examples: "error_count", "nginx/requests".

Metric identifiers are limited to 100 characters and can include only the following characters: A-Z, a-z, 0-9, and the special characters _-.,+!*',()%/. The forward-slash character (/) denotes a hierarchy of name pieces, and it cannot be the first character of the name.

This field is the \[METRIC_ID\] part of a metric resource name in the format “projects/[PROJECT_ID]/metrics/[METRIC_ID]”. Example: If the resource name of a metric is "projects/my-project/metrics/nginx%2Frequests", this field’s value is "nginx/requests".

description: String

Optional. A description of this metric, which is used in documentation. The maximum length of the description is 8000 characters.

filter: String

Required. An advanced logs filter which is used to match log entries. Example:

 "resource.type=gae_app AND severity>=ERROR"

The maximum length of the filter is 20000 characters.

bucket_name: String

Optional. The resource name of the Log Bucket that owns the Log Metric. Only Log Buckets in projects are supported. The bucket has to be in the same project as the metric.

For example:

projects/my-project/locations/global/buckets/my-bucket

If empty, then the Log Metric is considered a non-Bucket Log Metric.

disabled: bool

Optional. If set to True, then this metric is disabled and it does not generate any points.

metric_descriptor: Option<MetricDescriptor>

Optional. The metric descriptor associated with the logs-based metric. If unspecified, it uses a default metric descriptor with a DELTA metric kind, INT64 value type, with no labels and a unit of “1”. Such a metric counts the number of log entries matching the filter expression.

The name, type, and description fields in the metric_descriptor are output only, and is constructed using the name and description field in the LogMetric.

To create a logs-based metric that records a distribution of log values, a DELTA metric kind with a DISTRIBUTION value type must be used along with a value_extractor expression in the LogMetric.

Each label in the metric descriptor must have a matching label name as the key and an extractor expression as the value in the label_extractors map.

The metric_kind and value_type fields in the metric_descriptor cannot be updated once initially configured. New labels can be added in the metric_descriptor, but existing labels cannot be modified except for their description.

value_extractor: String

Optional. A value_extractor is required when using a distribution logs-based metric to extract the values to record from a log entry. Two functions are supported for value extraction: EXTRACT(field) or REGEXP_EXTRACT(field, regex). The arguments are:

field: The name of the log entry field from which the value is to be extracted.

regex: A regular expression using the Google RE2 syntax (https://github.com/google/re2/wiki/Syntax) with a single capture group to extract data from the specified log entry field. The value of the field is converted to a string before applying the regex. It is an error to specify a regex that does not include exactly one capture group.

The result of the extraction must be convertible to a double type, as the distribution always records double values. If either the extraction or the conversion to double fails, then those values are not recorded in the distribution.

Example: REGEXP_EXTRACT(jsonPayload.request, ".*quantity=(\d+).*")

label_extractors: BTreeMap<String, String>

Optional. A map from a label key string to an extractor expression which is used to extract data from a log entry field and assign as the label value. Each label key specified in the LabelDescriptor must have an associated extractor expression in this map. The syntax of the extractor expression is the same as for the value_extractor field.

The extracted value is converted to the type defined in the label descriptor. If either the extraction or the type conversion fails, the label will have a default value. The default value for a string label is an empty string, for an integer label its 0, and for a boolean label its false.

Note that there are upper bounds on the maximum number of labels and the number of active time series that are allowed in a project.

bucket_options: Option<BucketOptions>

Optional. The bucket_options are required when the logs-based metric is using a DISTRIBUTION value type and it describes the bucket boundaries used to create a histogram of the extracted values.

create_time: Option<Timestamp>

Output only. The creation timestamp of the metric.

This field may not be present for older metrics.

update_time: Option<Timestamp>

Output only. The last update timestamp of the metric.

This field may not be present for older metrics.

version: i32
👎Deprecated

Deprecated. The API version that created or updated this metric. The v2 format is used by default and cannot be changed.

Implementations

impl LogMetric

pub fn version(&self) -> ApiVersion

Returns the enum value of version, or the default if the field is set to an invalid enum value.

pub fn set_version(&mut self, value: ApiVersion)

Sets version to the provided enum value.