Module google_api_proto::google::maps::addressvalidation::v1

Modules

address_component

Nested message and enum types in AddressComponent.

address_validation_client

Generated client implementations.

provide_validation_feedback_request

Nested message and enum types in ProvideValidationFeedbackRequest.

verdict

Nested message and enum types in Verdict.

Structs

Address

Details of the post-processed address. Post-processing includes correcting misspelled parts of the address, replacing incorrect parts, and inferring missing parts.

AddressComponent

Represents an address component, such as a street, city, or state.

AddressMetadata

The metadata for the address. metadata is not guaranteed to be fully populated for every address sent to the Address Validation API.

ComponentName

A wrapper for the name of the component.

Geocode

Contains information about the place the input was geocoded to.

PlusCode

Plus code (http://plus.codes) is a location reference with two formats: global code defining a 14mx14m (1/8000th of a degree) or smaller rectangle, and compound code, replacing the prefix with a reference location.

ProvideValidationFeedbackRequest

The request for sending validation feedback.

ProvideValidationFeedbackResponse

The response for validation feedback.

UspsAddress

USPS representation of a US address.

UspsData

The USPS data for the address. uspsData is not guaranteed to be fully populated for every US or PR address sent to the Address Validation API. It’s recommended to integrate the backup address fields in the response if you utilize uspsData as the primary part of the response.

ValidateAddressRequest

The request for validating an address.

ValidateAddressResponse

The response to an address validation request.

ValidationResult

The result of validating an address.

Verdict

High level overview of the address validation result and geocode.