Module google_api_proto::google::maps::mapsplatformdatasets::v1alpha

Modules

dataset

Nested message and enum types in Dataset.

maps_platform_datasets_v1_alpha_client

Generated client implementations.

Structs

CreateDatasetRequest

Request to create a maps dataset.

Dataset

A representation of a maps platform dataset.

DeleteDatasetRequest

Request to delete a dataset.

DeleteDatasetVersionRequest

Request to delete a version of a dataset.

GcsSource

The details about the data source when it is in Google Cloud Storage.

GetDatasetRequest

Request to get the specified dataset.

ListDatasetVersionsRequest

Request to list of all versions of the dataset.

ListDatasetVersionsResponse

Response with list of all versions of the dataset.

ListDatasetsRequest

Request to list datasets for the project.

ListDatasetsResponse

Response to list datasets for the project.

LocalFileSource

The details about the data source when it is a local file.

UpdateDatasetMetadataRequest

Request to update the metadata fields of the dataset.

Enums

FileFormat

The format of the file being uploaded.

State

State specifies the status of the import of the latest dataset version.

Usage

Usage specifies where the data is intended to be used to inform how to process the data.