Module google_api_proto::google::maps::places::v1

Modules

autocomplete_places_request

Nested message and enum types in AutocompletePlacesRequest.

autocomplete_places_response

Nested message and enum types in AutocompletePlacesResponse.

contextual_content

Nested message and enum types in ContextualContent.

ev_charge_options

Nested message and enum types in EVChargeOptions.

fuel_options

Nested message and enum types in FuelOptions.

place

Nested message and enum types in Place.

places_client

Generated client implementations.

search_nearby_request

Nested message and enum types in SearchNearbyRequest.

search_text_request

Nested message and enum types in SearchTextRequest.

Structs

AuthorAttribution

Information about the author of the UGC data. Used in [Photo][google.maps.places.v1.Photo], and [Review][google.maps.places.v1.Review].

AutocompletePlacesRequest

Request proto for AutocompletePlaces.

AutocompletePlacesResponse

Response proto for AutocompletePlaces.

Circle

Circle with a LatLng as center and radius.

ContentBlock

A block of content that can be served individually.

ContextualContent

Experimental: See https://developers.google.com/maps/documentation/places/web-service/experimental/places-generative for more details.

EvChargeOptions

Information about the EV Charge Station hosted in Place. Terminology follows https://afdc.energy.gov/fuels/electricity_infrastructure.html One port could charge one car at a time. One port has one or more connectors. One station has one or more ports.

FuelOptions

The most recent information about fuel options in a gas station. This information is updated regularly.

GetPhotoMediaRequest

Request for fetching a photo of a place using a photo resource name.

GetPlaceRequest

Request for fetching a Place based on its resource name, which is a string in the places/{place_id} format.

Photo

Information about a photo of a place.

PhotoMedia

A photo media from Places API.

Place

All the information representing a Place.

References

Experimental: See https://developers.google.com/maps/documentation/places/web-service/experimental/places-generative for more details.

Review

Information about a review of a place.

SearchNearbyRequest

Request proto for Search Nearby.

SearchNearbyResponse

Response proto for Search Nearby.

SearchTextRequest

Request proto for SearchText.

SearchTextResponse

Response proto for SearchText.

Enums

EvConnectorType

See http://ieeexplore.ieee.org/stamp/stamp.jsp?arnumber=6872107 for additional information/context on EV charging connector types.

PriceLevel

Price level of the place.