Module google_api_proto::google::maps::places::v1::autocomplete_places_response::suggestion

Nested message and enum types in Suggestion.

Structs

FormattableText

Text representing a Place or query prediction. The text may be used as is or formatted.

PlacePrediction

Prediction results for a Place Autocomplete prediction.

QueryPrediction

Prediction results for a Query Autocomplete prediction.

StringRange

Identifies a substring within a given text.

StructuredFormat

Contains a breakdown of a Place or query prediction into main text and secondary text.

Enums

Kind