Module google_api_proto::google::maps::playablelocations::v3::sample

Modules

playable_location

Nested message and enum types in PlayableLocation.

spacing_options

Nested message and enum types in SpacingOptions.

Structs

AreaFilter

Specifies the area to search for playable locations.

Criterion

Encapsulates a filter criterion for searching for a set of playable locations.

Filter

Specifies the filters to use when searching for playable locations.

PlayableLocation

A geographical point suitable for placing game objects in location-based games.

PlayableLocationList

A list of PlayableLocation objects that satisfies a single Criterion.

SpacingOptions

A set of options that specifies the separation between playable locations.