Module google_api_proto::google::maps::solar::v1

Modules

solar_client

Generated client implementations.

Structs

BuildingInsights

Response message for Solar.FindClosestBuildingInsights. Information about the location, dimensions, and solar potential of a building.

CashPurchaseSavings

Cost and benefit of an outright purchase of a particular configuration of solar panels with a particular electricity usage.

DataLayers

Information about the solar potential of a region. The actual data are contained in a number of GeoTIFF files covering the requested region, for which this message contains URLs: Each string in the DataLayers message contains a URL from which the corresponding GeoTIFF can be fetched. These URLs are valid for a few hours after they’ve been generated. Most of the GeoTIFF files are at a resolution of 0.1m/pixel, but the monthly flux file is at 0.5m/pixel, and the hourly shade files are at 1m/pixel. If a pixel_size_meters value was specified in the GetDataLayersRequest, then the minimum resolution in the GeoTIFF files will be that value.

FinancedPurchaseSavings

Cost and benefit of using a loan to buy a particular configuration of solar panels with a particular electricity usage.

FinancialAnalysis

Analysis of the cost and benefits of the optimum solar layout for a particular electric bill size.

FinancialDetails

Details of a financial analysis. Some of these details are already stored at higher levels (e.g., out of pocket cost). Total money amounts are over a lifetime period defined by the [panel_lifetime_years] [google.maps.solar.v1.SolarPotential.panel_lifetime_years] field in [SolarPotential] [google.maps.solar.v1.SolarPotential]. Note: The out of pocket cost of purchasing the panels is given in the [out_of_pocket_cost] [google.maps.solar.v1.CashPurchaseSavings.out_of_pocket_cost] field in [CashPurchaseSavings] [google.maps.solar.v1.CashPurchaseSavings].

FindClosestBuildingInsightsRequest

Request message for Solar.FindClosestBuildingInsights.

GetDataLayersRequest

Request message for Solar.GetDataLayers.

GetGeoTiffRequest

Request message for Solar.GetGeoTiff.

LatLngBox

A bounding box in lat/lng coordinates.

LeasingSavings

Cost and benefit of leasing a particular configuration of solar panels with a particular electricity usage.

RoofSegmentSizeAndSunshineStats

Information about the size and sunniness quantiles of a roof segment.

RoofSegmentSummary

Information about a roof segment on the building, with some number of panels placed on it.

SavingsOverTime

Financial information that’s shared between different financing methods.

SizeAndSunshineStats

Size and sunniness quantiles of a roof, or part of a roof.

SolarPanel

SolarPanel describes the position, orientation, and production of a single solar panel. See the [panel_height_meters] [google.maps.solar.v1.SolarPotential.panel_height_meters], [panel_width_meters] [google.maps.solar.v1.SolarPotential.panel_width_meters], and [panel_capacity_watts] [google.maps.solar.v1.SolarPotential.panel_capacity_watts] fields in [SolarPotential] [google.maps.solar.v1.SolarPotential] for information on the parameters of the panel.

SolarPanelConfig

SolarPanelConfig describes a particular placement of solar panels on the roof.

SolarPotential

Information about the solar potential of a building. A number of fields in this are defined in terms of “panels”. The fields [panel_capacity_watts] [google.maps.solar.v1.SolarPotential.panel_capacity_watts], [panel_height_meters] [google.maps.solar.v1.SolarPotential.panel_height_meters], and [panel_width_meters] [google.maps.solar.v1.SolarPotential.panel_width_meters] describe the parameters of the model of panel used in these calculations.

Enums

DataLayerView

What subset of the solar information to return.

ImageryQuality

The quality of the imagery used to compute some API result.

SolarPanelOrientation

The orientation of a solar panel. This must be interpreted relative to the azimuth of the roof segment that the panel is placed on.