Struct google_api_proto::google::maps::solar::v1::SolarPotential

pub struct SolarPotential {

    pub max_array_panels_count: i32,
    pub panel_capacity_watts: f32,
    pub panel_height_meters: f32,
    pub panel_width_meters: f32,
    pub panel_lifetime_years: i32,
    pub max_array_area_meters2: f32,
    pub max_sunshine_hours_per_year: f32,
    pub carbon_offset_factor_kg_per_mwh: f32,
    pub whole_roof_stats: Option<SizeAndSunshineStats>,
    pub building_stats: Option<SizeAndSunshineStats>,
    pub roof_segment_stats: Vec<RoofSegmentSizeAndSunshineStats>,
    pub solar_panels: Vec<SolarPanel>,
    pub solar_panel_configs: Vec<SolarPanelConfig>,
    pub financial_analyses: Vec<FinancialAnalysis>,

}

Information about the solar potential of a building. A number of fields in this are defined in terms of “panels”. The fields [panel_capacity_watts] [google.maps.solar.v1.SolarPotential.panel_capacity_watts], [panel_height_meters] [google.maps.solar.v1.SolarPotential.panel_height_meters], and [panel_width_meters] [google.maps.solar.v1.SolarPotential.panel_width_meters] describe the parameters of the model of panel used in these calculations.

Fields
max_array_panels_count: i32

Size of the maximum array - that is, the maximum number of panels that can fit on the roof.

panel_capacity_watts: f32

Capacity, in watts, of the panel used in the calculations.

panel_height_meters: f32

Height, in meters in portrait orientation, of the panel used in the calculations.

panel_width_meters: f32

Width, in meters in portrait orientation, of the panel used in the calculations.

panel_lifetime_years: i32

The expected lifetime, in years, of the solar panels. This is used in the financial calculations.

max_array_area_meters2: f32

Size, in square meters, of the maximum array.

max_sunshine_hours_per_year: f32

Maximum number of sunshine hours received per year, by any point on the roof. Sunshine hours are a measure of the total amount of insolation (energy) received per year. 1 sunshine hour = 1 kWh per kW (where kW refers to kW of capacity under Standard Testing Conditions).

carbon_offset_factor_kg_per_mwh: f32

Equivalent amount of CO2 produced per MWh of grid electricity. This is a measure of the carbon intensity of grid electricity displaced by solar electricity.

whole_roof_stats: Option<SizeAndSunshineStats>

Total size and sunlight quantiles for the part of the roof that was assigned to some roof segment. Despite the name, this may not include the entire building. See [building_stats] [google.maps.solar.v1.SolarPotential.building_stats].

building_stats: Option<SizeAndSunshineStats>

Size and sunlight quantiles for the entire building, including parts of the roof that were not assigned to some roof segment. Because the orientations of these parts are not well characterised, the roof area estimate is unreliable, but the ground area estimate is reliable. It may be that a more reliable whole building roof area can be obtained by scaling the roof area from [whole_roof_stats] [google.maps.solar.v1.SolarPotential.whole_roof_stats] by the ratio of the ground areas of building_stats and whole_roof_stats.

roof_segment_stats: Vec<RoofSegmentSizeAndSunshineStats>

Size and sunlight quantiles for each roof segment.

solar_panels: Vec<SolarPanel>

Each [SolarPanel] [google.maps.solar.v1.SolarPanel] describes a single solar panel. They are listed in the order that the panel layout algorithm placed this. This is usually, though not always, in decreasing order of annual energy production.

solar_panel_configs: Vec<SolarPanelConfig>

Each [SolarPanelConfig] [google.maps.solar.v1.SolarPanelConfig] describes a different arrangement of solar panels on the roof. They are in order of increasing number of panels. The SolarPanelConfig with [panels_count] [google.maps.solar.v1.SolarPanelConfig.panels_count]=N is based on the first N panels in the solar_panels list.

financial_analyses: Vec<FinancialAnalysis>

A [FinancialAnalysis] [google.maps.solar.v1.FinancialAnalysis] gives the savings from going solar assuming a given monthly bill and a given electricity provider. They are in order of increasing order of monthly bill amount. This field will be empty for buildings in areas for which the Solar API does not have enough information to perform financial computations.