Module google_api_proto::google::marketingplatform::admin::v1alpha

Modules

marketingplatform_admin_service_client

Generated client implementations.

Structs

AnalyticsAccountLink

A resource message representing the link between a Google Analytics account and a Google Marketing Platform organization.

CreateAnalyticsAccountLinkRequest

Request message for CreateAnalyticsAccountLink RPC.

DeleteAnalyticsAccountLinkRequest

Request message for DeleteAnalyticsAccountLink RPC.

GetOrganizationRequest

Request message for GetOrganization RPC.

ListAnalyticsAccountLinksRequest

Request message for ListAnalyticsAccountLinks RPC.

ListAnalyticsAccountLinksResponse

Response message for ListAnalyticsAccountLinks RPC.

Organization

A resource message representing a Google Marketing Platform organization.

SetPropertyServiceLevelRequest

Request message for SetPropertyServiceLevel RPC.

SetPropertyServiceLevelResponse

Response message for SetPropertyServiceLevel RPC.

Enums

AnalyticsServiceLevel

Various levels of service for Google Analytics.

LinkVerificationState

The verification state of the link between a product account and a GMP organization.