Struct google_api_proto::google::monitoring::v3::metric_service_client::MetricServiceClient

pub struct MetricServiceClient<T> { /* private fields */ }

Manages metric descriptors, monitored resource descriptors, and time series data.

Implementations

impl<T> MetricServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> MetricServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_monitored_resource_descriptors(
&mut self,
request: impl IntoRequest<ListMonitoredResourceDescriptorsRequest>
) -> Result<Response<ListMonitoredResourceDescriptorsResponse>, Status>

Lists monitored resource descriptors that match a filter. This method does not require a Workspace.

pub async fn get_monitored_resource_descriptor(
&mut self,
request: impl IntoRequest<GetMonitoredResourceDescriptorRequest>
) -> Result<Response<MonitoredResourceDescriptor>, Status>

Gets a single monitored resource descriptor. This method does not require a Workspace.

pub async fn list_metric_descriptors(
&mut self,
request: impl IntoRequest<ListMetricDescriptorsRequest>
) -> Result<Response<ListMetricDescriptorsResponse>, Status>

Lists metric descriptors that match a filter. This method does not require a Workspace.

pub async fn get_metric_descriptor(
&mut self,
request: impl IntoRequest<GetMetricDescriptorRequest>
) -> Result<Response<MetricDescriptor>, Status>

Gets a single metric descriptor. This method does not require a Workspace.

pub async fn create_metric_descriptor(
&mut self,
request: impl IntoRequest<CreateMetricDescriptorRequest>
) -> Result<Response<MetricDescriptor>, Status>

Creates a new metric descriptor. The creation is executed asynchronously and callers may check the returned operation to track its progress. User-created metric descriptors define custom metrics.

pub async fn delete_metric_descriptor(
&mut self,
request: impl IntoRequest<DeleteMetricDescriptorRequest>
) -> Result<Response<()>, Status>

Deletes a metric descriptor. Only user-created custom metrics can be deleted.

pub async fn list_time_series(
&mut self,
request: impl IntoRequest<ListTimeSeriesRequest>
) -> Result<Response<ListTimeSeriesResponse>, Status>

Lists time series that match a filter. This method does not require a Workspace.

pub async fn create_time_series(
&mut self,
request: impl IntoRequest<CreateTimeSeriesRequest>
) -> Result<Response<()>, Status>

Creates or adds data to one or more time series. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response.

pub async fn create_service_time_series(
&mut self,
request: impl IntoRequest<CreateTimeSeriesRequest>
) -> Result<Response<()>, Status>

Creates or adds data to one or more service time series. A service time series is a time series for a metric from a Google Cloud service. The response is empty if all time series in the request were written. If any time series could not be written, a corresponding failure message is included in the error response. This endpoint rejects writes to user-defined metrics. This method is only for use by Google Cloud services. Use [projects.timeSeries.create][google.monitoring.v3.MetricService.CreateTimeSeries] instead.