Module google_api_proto::google::monitoring::v3::service

Nested message and enum types in Service.

Structs

AppEngine

App Engine service. Learn more at https://cloud.google.com/appengine.

BasicService

A well-known service type, defined by its service type and service labels. Documentation and examples here.

CloudEndpoints

Cloud Endpoints service. Learn more at https://cloud.google.com/endpoints.

CloudRun

Cloud Run service. Learn more at https://cloud.google.com/run.

ClusterIstio

Istio service scoped to a single Kubernetes cluster. Learn more at https://istio.io. Clusters running OSS Istio will have their services ingested as this type.

Custom

Use a custom service to designate a service that you want to monitor when none of the other service types (like App Engine, Cloud Run, or a GKE type) matches your intended service.

GkeNamespace

GKE Namespace. The field names correspond to the resource metadata labels on monitored resources that fall under a namespace (for example, k8s_container or k8s_pod).

GkeService

GKE Service. The “service” here represents a Kubernetes service object. The field names correspond to the resource labels on k8s_service monitored resources.

GkeWorkload

A GKE Workload (Deployment, StatefulSet, etc). The field names correspond to the metadata labels on monitored resources that fall under a workload (for example, k8s_container or k8s_pod).

IstioCanonicalService

Canonical service scoped to an Istio mesh. Anthos clusters running ASM >= 1.6.8 will have their services ingested as this type.

MeshIstio

Istio service scoped to an Istio mesh. Anthos clusters running ASM < 1.6.8 will have their services ingested as this type.

Telemetry

Configuration for how to query telemetry on a Service.

Enums

Identifier

REQUIRED. Service-identifying atoms specifying the underlying service.