Module google_api_proto::google::monitoring::v3::uptime_check_config::http_check

Nested message and enum types in HttpCheck.

Structs

BasicAuthentication

The authentication parameters to provide to the specified resource or URL that requires a username and password. Currently, only Basic HTTP authentication is supported in Uptime checks.

Enums

ContentType

Header options corresponding to the content type of a HTTP request body.

RequestMethod

The HTTP request method options.