Module google_api_proto::google::privacy::dlp::v2::action

Nested message and enum types in Action.

Modules

deidentify

Nested message and enum types in Deidentify.

Structs

Deidentify

Create a de-identified copy of the requested table or files.

JobNotificationEmails

Sends an email when the job completes. The email goes to IAM project owners and technical Essential Contacts.

PublishFindingsToCloudDataCatalog

Publish findings of a DlpJob to Data Catalog. In Data Catalog, tag templates are applied to the resource that Cloud DLP scanned. Data Catalog tag templates are stored in the same project and region where the BigQuery table exists. For Cloud DLP to create and apply the tag template, the Cloud DLP service agent must have the roles/datacatalog.tagTemplateOwner permission on the project. The tag template contains fields summarizing the results of the DlpJob. Any field values previously written by another DlpJob are deleted. [InfoType naming patterns][google.privacy.dlp.v2.InfoType] are strictly enforced when using this feature.

PublishSummaryToCscc

Publish the result summary of a DlpJob to Security Command Center. This action is available for only projects that belong to an organization. This action publishes the count of finding instances and their infoTypes. The summary of findings are persisted in Security Command Center and are governed by service-specific policies for Security Command Center. Only a single instance of this action can be specified. Compatible with: Inspect

PublishToPubSub

Publish a message into a given Pub/Sub topic when DlpJob has completed. The message contains a single field, DlpJobName, which is equal to the finished job’s DlpJob.name. Compatible with: Inspect, Risk

PublishToStackdriver

Enable Stackdriver metric dlp.googleapis.com/finding_count. This will publish a metric to stack driver on each infotype requested and how many findings were found for it. CustomDetectors will be bucketed as ‘Custom’ under the Stackdriver label ‘info_type’.

SaveFindings

If set, the detailed findings will be persisted to the specified OutputStorageConfig. Only a single instance of this action can be specified. Compatible with: Inspect, Risk

Enums

Action

Extra events to execute after the job has finished.