Module google_api_proto::google::privacy::dlp::v2::analyze_data_source_risk_details

Nested message and enum types in AnalyzeDataSourceRiskDetails.

Modules

categorical_stats_result

Nested message and enum types in CategoricalStatsResult.

delta_presence_estimation_result

Nested message and enum types in DeltaPresenceEstimationResult.

k_anonymity_result

Nested message and enum types in KAnonymityResult.

k_map_estimation_result

Nested message and enum types in KMapEstimationResult.

l_diversity_result

Nested message and enum types in LDiversityResult.

Structs

CategoricalStatsResult

Result of the categorical stats computation.

DeltaPresenceEstimationResult

Result of the δ-presence computation. Note that these results are an estimation, not exact values.

KAnonymityResult

Result of the k-anonymity computation.

KMapEstimationResult

Result of the reidentifiability analysis. Note that these results are an estimation, not exact values.

LDiversityResult

Result of the l-diversity computation.

NumericalStatsResult

Result of the numerical stats computation.

RequestedRiskAnalysisOptions

Risk analysis options.

Enums

Result

Values associated with this metric.