Module google_api_proto::google::privacy::dlp::v2::custom_info_type

Nested message and enum types in CustomInfoType.

Modules

detection_rule

Nested message and enum types in DetectionRule.

dictionary

Nested message and enum types in Dictionary.

Structs

DetectionRule

Deprecated; use InspectionRuleSet instead. Rule for modifying a CustomInfoType to alter behavior under certain circumstances, depending on the specific details of the rule. Not supported for the surrogate_type custom infoType.

Dictionary

Custom information type based on a dictionary of words or phrases. This can be used to match sensitive information specific to the data, such as a list of employee IDs or job titles.

Regex

Message defining a custom regular expression.

SurrogateType

Message for detecting output from deidentification transformations such as CryptoReplaceFfxFpeConfig. These types of transformations are those that perform pseudonymization, thereby producing a “surrogate” as output. This should be used in conjunction with a field on the transformation such as surrogate_info_type. This CustomInfoType does not support the use of detection_rules.

Enums

ExclusionType

Type of exclusion rule.

Type

Type of custom detector.