Enum google_api_proto::google::privacy::dlp::v2::TransformationType

#[repr(i32)]
pub enum TransformationType {

    Unspecified,
    RecordSuppression,
    ReplaceValue,
    ReplaceDictionary,
    Redact,
    CharacterMask,
    CryptoReplaceFfxFpe,
    FixedSizeBucketing,
    Bucketing,
    ReplaceWithInfoType,
    TimePart,
    CryptoHash,
    DateShift,
    CryptoDeterministicConfig,
    RedactImage,

}

An enum of rules that can be used to transform a value. Can be a record suppression, or one of the transformation rules specified under PrimitiveTransformation.

Variants

Unspecified

Unused

RecordSuppression

Record suppression

ReplaceValue

Replace value

ReplaceDictionary

Replace value using a dictionary.

Redact

Redact

CharacterMask

Character mask

CryptoReplaceFfxFpe

FFX-FPE

FixedSizeBucketing

Fixed size bucketing

Bucketing

Bucketing

ReplaceWithInfoType

Replace with info type

TimePart

Time part

CryptoHash

Crypto hash

DateShift

Date shift

CryptoDeterministicConfig

Deterministic crypto

RedactImage

Redact image

Implementations

impl TransformationType

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of TransformationType.

pub fn from_i32(value: i32) -> Option<TransformationType>

Converts an i32 to a TransformationType, or None if value is not a valid variant.

impl TransformationType

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.