Module google_api_proto::google::privacy::dlp::v2::record_condition

Nested message and enum types in RecordCondition.

Modules

expressions

Nested message and enum types in Expressions.

Structs

Condition

The field type of value and field do not need to match to be considered equal, but not all comparisons are possible. EQUAL_TO and NOT_EQUAL_TO attempt to compare even with incompatible types, but all other comparisons are invalid with incompatible types. A value of type:

Conditions

A collection of conditions.

Expressions

An expression, consisting or an operator and conditions.