Module google_api_proto::google::pubsub::v1::push_config

Nested message and enum types in PushConfig.

Structs

NoWrapper

Sets the data field as the HTTP body for delivery.

OidcToken

Contains information needed for generating an OpenID Connect token.

PubsubWrapper

The payload to the push endpoint is in the form of the JSON representation of a PubsubMessage (https://cloud.google.com/pubsub/docs/reference/rpc/google.pubsub.v1#pubsubmessage).

Enums

AuthenticationMethod

An authentication method used by push endpoints to verify the source of push requests. This can be used with push endpoints that are private by default to allow requests only from the Pub/Sub system, for example. This field is optional and should be set only by users interested in authenticated push.

Wrapper

The format of the delivered message to the push endpoint is defined by the chosen wrapper. When unset, PubsubWrapper is used.