Struct google_api_proto::google::pubsub::v1::subscriber_client::SubscriberClient

pub struct SubscriberClient<T> { /* private fields */ }

The service that an application uses to manipulate subscriptions and to consume messages from a subscription via the Pull method or by establishing a bi-directional stream using the StreamingPull method.

Implementations

impl<T> SubscriberClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> SubscriberClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_subscription( &mut self, request: impl IntoRequest<Subscription>, ) -> Result<Response<Subscription>, Status>

Creates a subscription to a given topic. See the [resource name rules] (https://cloud.google.com/pubsub/docs/pubsub-basics#resource_names). If the subscription already exists, returns ALREADY_EXISTS. If the corresponding topic doesn’t exist, returns NOT_FOUND.

If the name is not provided in the request, the server will assign a random name for this subscription on the same project as the topic, conforming to the [resource name format] (https://cloud.google.com/pubsub/docs/pubsub-basics#resource_names). The generated name is populated in the returned Subscription object. Note that for REST API requests, you must specify a name in the request.

pub async fn get_subscription( &mut self, request: impl IntoRequest<GetSubscriptionRequest>, ) -> Result<Response<Subscription>, Status>

Gets the configuration details of a subscription.

pub async fn update_subscription( &mut self, request: impl IntoRequest<UpdateSubscriptionRequest>, ) -> Result<Response<Subscription>, Status>

Updates an existing subscription by updating the fields specified in the update mask. Note that certain properties of a subscription, such as its topic, are not modifiable.

pub async fn list_subscriptions( &mut self, request: impl IntoRequest<ListSubscriptionsRequest>, ) -> Result<Response<ListSubscriptionsResponse>, Status>

Lists matching subscriptions.

pub async fn delete_subscription( &mut self, request: impl IntoRequest<DeleteSubscriptionRequest>, ) -> Result<Response<()>, Status>

Deletes an existing subscription. All messages retained in the subscription are immediately dropped. Calls to Pull after deletion will return NOT_FOUND. After a subscription is deleted, a new one may be created with the same name, but the new one has no association with the old subscription or its topic unless the same topic is specified.

pub async fn modify_ack_deadline( &mut self, request: impl IntoRequest<ModifyAckDeadlineRequest>, ) -> Result<Response<()>, Status>

Modifies the ack deadline for a specific message. This method is useful to indicate that more time is needed to process a message by the subscriber, or to make the message available for redelivery if the processing was interrupted. Note that this does not modify the subscription-level ackDeadlineSeconds used for subsequent messages.

pub async fn acknowledge( &mut self, request: impl IntoRequest<AcknowledgeRequest>, ) -> Result<Response<()>, Status>

Acknowledges the messages associated with the ack_ids in the AcknowledgeRequest. The Pub/Sub system can remove the relevant messages from the subscription.

Acknowledging a message whose ack deadline has expired may succeed, but such a message may be redelivered later. Acknowledging a message more than once will not result in an error.

pub async fn pull( &mut self, request: impl IntoRequest<PullRequest>, ) -> Result<Response<PullResponse>, Status>

Pulls messages from the server.

pub async fn streaming_pull( &mut self, request: impl IntoStreamingRequest<Message = StreamingPullRequest>, ) -> Result<Response<Streaming<StreamingPullResponse>>, Status>

Establishes a stream with the server, which sends messages down to the client. The client streams acknowledgements and ack deadline modifications back to the server. The server will close the stream and return the status on any error. The server may close the stream with status UNAVAILABLE to reassign server-side resources, in which case, the client should re-establish the stream. Flow control can be achieved by configuring the underlying RPC channel.

pub async fn modify_push_config( &mut self, request: impl IntoRequest<ModifyPushConfigRequest>, ) -> Result<Response<()>, Status>

Modifies the PushConfig for a specified subscription.

This may be used to change a push subscription to a pull one (signified by an empty PushConfig) or vice versa, or change the endpoint URL and other attributes of a push subscription. Messages will accumulate for delivery continuously through the call regardless of changes to the PushConfig.

pub async fn get_snapshot( &mut self, request: impl IntoRequest<GetSnapshotRequest>, ) -> Result<Response<Snapshot>, Status>

Gets the configuration details of a snapshot. Snapshots are used in Seek operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot.

pub async fn list_snapshots( &mut self, request: impl IntoRequest<ListSnapshotsRequest>, ) -> Result<Response<ListSnapshotsResponse>, Status>

Lists the existing snapshots. Snapshots are used in Seek operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot.

pub async fn create_snapshot( &mut self, request: impl IntoRequest<CreateSnapshotRequest>, ) -> Result<Response<Snapshot>, Status>

Creates a snapshot from the requested subscription. Snapshots are used in Seek operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot. If the snapshot already exists, returns ALREADY_EXISTS. If the requested subscription doesn’t exist, returns NOT_FOUND. If the backlog in the subscription is too old – and the resulting snapshot would expire in less than 1 hour – then FAILED_PRECONDITION is returned. See also the Snapshot.expire_time field. If the name is not provided in the request, the server will assign a random name for this snapshot on the same project as the subscription, conforming to the [resource name format] (https://cloud.google.com/pubsub/docs/pubsub-basics#resource_names). The generated name is populated in the returned Snapshot object. Note that for REST API requests, you must specify a name in the request.

pub async fn update_snapshot( &mut self, request: impl IntoRequest<UpdateSnapshotRequest>, ) -> Result<Response<Snapshot>, Status>

Updates an existing snapshot by updating the fields specified in the update mask. Snapshots are used in Seek operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot.

pub async fn delete_snapshot( &mut self, request: impl IntoRequest<DeleteSnapshotRequest>, ) -> Result<Response<()>, Status>

Removes an existing snapshot. Snapshots are used in [Seek] (https://cloud.google.com/pubsub/docs/replay-overview) operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot. When the snapshot is deleted, all messages retained in the snapshot are immediately dropped. After a snapshot is deleted, a new one may be created with the same name, but the new one has no association with the old snapshot or its subscription, unless the same subscription is specified.

pub async fn seek( &mut self, request: impl IntoRequest<SeekRequest>, ) -> Result<Response<SeekResponse>, Status>

Seeks an existing subscription to a point in time or to a given snapshot, whichever is provided in the request. Snapshots are used in [Seek] (https://cloud.google.com/pubsub/docs/replay-overview) operations, which allow you to manage message acknowledgments in bulk. That is, you can set the acknowledgment state of messages in an existing subscription to the state captured by a snapshot. Note that both the subscription and the snapshot must be on the same topic.