Struct google_api_proto::google::shopping::css::v1::Attributes

pub struct Attributes {

    pub cpp_link: Option<String>,
    pub cpp_mobile_link: Option<String>,
    pub cpp_ads_redirect: Option<String>,
    pub low_price: Option<Price>,
    pub high_price: Option<Price>,
    pub number_of_offers: Option<i64>,
    pub headline_offer_condition: Option<String>,
    pub headline_offer_price: Option<Price>,
    pub headline_offer_link: Option<String>,
    pub headline_offer_mobile_link: Option<String>,
    pub headline_offer_shipping_price: Option<Price>,
    pub title: Option<String>,
    pub image_link: Option<String>,
    pub additional_image_links: Vec<String>,
    pub description: Option<String>,
    pub brand: Option<String>,
    pub mpn: Option<String>,
    pub gtin: Option<String>,
    pub product_types: Vec<String>,
    pub google_product_category: Option<String>,
    pub adult: Option<bool>,
    pub multipack: Option<i64>,
    pub is_bundle: Option<bool>,
    pub age_group: Option<String>,
    pub color: Option<String>,
    pub gender: Option<String>,
    pub material: Option<String>,
    pub pattern: Option<String>,
    pub size: Option<String>,
    pub size_system: Option<String>,
    pub size_types: Vec<String>,
    pub item_group_id: Option<String>,
    pub product_details: Vec<ProductDetail>,
    pub product_weight: Option<ProductWeight>,
    pub product_length: Option<ProductDimension>,
    pub product_width: Option<ProductDimension>,
    pub product_height: Option<ProductDimension>,
    pub product_highlights: Vec<String>,
    pub certifications: Vec<Certification>,
    pub expiration_date: Option<Timestamp>,
    pub included_destinations: Vec<String>,
    pub excluded_destinations: Vec<String>,
    pub pause: Option<String>,
    pub custom_label_0: Option<String>,
    pub custom_label_1: Option<String>,
    pub custom_label_2: Option<String>,
    pub custom_label_3: Option<String>,
    pub custom_label_4: Option<String>,

}

Attributes for CSS Product.

Fields
cpp_link: Option<String>

URL directly linking to your the Product Detail Page of the CSS.

cpp_mobile_link: Option<String>

URL for the mobile-optimized version of the Product Detail Page of the CSS.

cpp_ads_redirect: Option<String>

Allows advertisers to override the item URL when the product is shown within the context of Product Ads.

low_price: Option<Price>

Low Price of the aggregate offer.

high_price: Option<Price>

High Price of the aggregate offer.

number_of_offers: Option<i64>

The number of aggregate offers.

headline_offer_condition: Option<String>

Condition of the headline offer.

headline_offer_price: Option<Price>

Headline Price of the aggregate offer.

headline_offer_link: Option<String>

Link to the headline offer.

headline_offer_mobile_link: Option<String>

Mobile Link to the headline offer.

headline_offer_shipping_price: Option<Price>

Headline Price of the aggregate offer.

title: Option<String>

Title of the item.

image_link: Option<String>

URL of an image of the item.

additional_image_links: Vec<String>

Additional URL of images of the item.

description: Option<String>

Description of the item.

brand: Option<String>

Product Related Attributes.[14-36] Brand of the item.

mpn: Option<String>

Manufacturer Part Number (MPN) of the item.

gtin: Option<String>

Global Trade Item Number (GTIN) of the item.

product_types: Vec<String>

Categories of the item (formatted as in products data specification).

google_product_category: Option<String>

Google’s category of the item (see Google product taxonomy). When querying products, this field will contain the user provided value. There is currently no way to get back the auto assigned google product categories through the API.

adult: Option<bool>

Set to true if the item is targeted towards adults.

multipack: Option<i64>

The number of identical products in a merchant-defined multipack.

is_bundle: Option<bool>

Whether the item is a merchant-defined bundle. A bundle is a custom grouping of different products sold by a merchant for a single price.

age_group: Option<String>

Target age group of the item.

color: Option<String>

Color of the item.

gender: Option<String>

Target gender of the item.

material: Option<String>

The material of which the item is made.

pattern: Option<String>

The item’s pattern (e.g. polka dots).

size: Option<String>

Size of the item. Only one value is allowed. For variants with different sizes, insert a separate product for each size with the same itemGroupId value (see [https://support.google.com/merchants/answer/6324492](size definition)).

size_system: Option<String>

System in which the size is specified. Recommended for apparel items.

size_types: Vec<String>

The cut of the item. It can be used to represent combined size types for apparel items. Maximum two of size types can be provided (see [https://support.google.com/merchants/answer/6324497](size type)).

item_group_id: Option<String>

Shared identifier for all variants of the same product.

product_details: Vec<ProductDetail>

Technical specification or additional product details.

product_weight: Option<ProductWeight>

The weight of the product in the units provided. The value must be between 0 (exclusive) and 2000 (inclusive).

product_length: Option<ProductDimension>

The length of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_width: Option<ProductDimension>

The width of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_height: Option<ProductDimension>

The height of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_highlights: Vec<String>

Bullet points describing the most relevant highlights of a product.

certifications: Vec<Certification>

A list of certificates claimed by the CSS for the given product.

expiration_date: Option<Timestamp>

Date on which the item should expire, as specified upon insertion, in ISO 8601 format. The actual expiration date is exposed in productstatuses as googleExpirationDate and might be earlier if expirationDate is too far in the future. Note: It may take 2+ days from the expiration date for the item to actually get deleted.

included_destinations: Vec<String>

The list of destinations to include for this target (corresponds to checked check boxes in Merchant Center). Default destinations are always included unless provided in excludedDestinations.

excluded_destinations: Vec<String>

The list of destinations to exclude for this target (corresponds to unchecked check boxes in Merchant Center).

pause: Option<String>

Publication of this item will be temporarily paused.

custom_label_0: Option<String>

Custom label 0 for custom grouping of items in a Shopping campaign.

custom_label_1: Option<String>

Custom label 1 for custom grouping of items in a Shopping campaign.

custom_label_2: Option<String>

Custom label 2 for custom grouping of items in a Shopping campaign.

custom_label_3: Option<String>

Custom label 3 for custom grouping of items in a Shopping campaign.

custom_label_4: Option<String>

Custom label 4 for custom grouping of items in a Shopping campaign.

Implementations

impl Attributes

pub fn cpp_link(&self) -> &str

Returns the value of cpp_link, or the default value if cpp_link is unset.

pub fn cpp_mobile_link(&self) -> &str

Returns the value of cpp_mobile_link, or the default value if cpp_mobile_link is unset.

pub fn number_of_offers(&self) -> i64

Returns the value of number_of_offers, or the default value if number_of_offers is unset.

pub fn headline_offer_condition(&self) -> &str

Returns the value of headline_offer_condition, or the default value if headline_offer_condition is unset.

pub fn headline_offer_link(&self) -> &str

Returns the value of headline_offer_link, or the default value if headline_offer_link is unset.

pub fn headline_offer_mobile_link(&self) -> &str

Returns the value of headline_offer_mobile_link, or the default value if headline_offer_mobile_link is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn image_link(&self) -> &str

Returns the value of image_link, or the default value if image_link is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn mpn(&self) -> &str

Returns the value of mpn, or the default value if mpn is unset.

pub fn gtin(&self) -> &str

Returns the value of gtin, or the default value if gtin is unset.

pub fn google_product_category(&self) -> &str

Returns the value of google_product_category, or the default value if google_product_category is unset.

pub fn adult(&self) -> bool

Returns the value of adult, or the default value if adult is unset.

pub fn multipack(&self) -> i64

Returns the value of multipack, or the default value if multipack is unset.

pub fn is_bundle(&self) -> bool

Returns the value of is_bundle, or the default value if is_bundle is unset.

pub fn age_group(&self) -> &str

Returns the value of age_group, or the default value if age_group is unset.

pub fn color(&self) -> &str

Returns the value of color, or the default value if color is unset.

pub fn gender(&self) -> &str

Returns the value of gender, or the default value if gender is unset.

pub fn material(&self) -> &str

Returns the value of material, or the default value if material is unset.

pub fn pattern(&self) -> &str

Returns the value of pattern, or the default value if pattern is unset.

pub fn size(&self) -> &str

Returns the value of size, or the default value if size is unset.

pub fn size_system(&self) -> &str

Returns the value of size_system, or the default value if size_system is unset.

pub fn item_group_id(&self) -> &str

Returns the value of item_group_id, or the default value if item_group_id is unset.

pub fn cpp_ads_redirect(&self) -> &str

Returns the value of cpp_ads_redirect, or the default value if cpp_ads_redirect is unset.

pub fn pause(&self) -> &str

Returns the value of pause, or the default value if pause is unset.

pub fn custom_label_0(&self) -> &str

Returns the value of custom_label_0, or the default value if custom_label_0 is unset.

pub fn custom_label_1(&self) -> &str

Returns the value of custom_label_1, or the default value if custom_label_1 is unset.

pub fn custom_label_2(&self) -> &str

Returns the value of custom_label_2, or the default value if custom_label_2 is unset.

pub fn custom_label_3(&self) -> &str

Returns the value of custom_label_3, or the default value if custom_label_3 is unset.

pub fn custom_label_4(&self) -> &str

Returns the value of custom_label_4, or the default value if custom_label_4 is unset.