Module google_api_proto::google::shopping::merchant::accounts::v1beta

Modules

account_issue

Nested message and enum types in AccountIssue.

account_issue_service_client

Generated client implementations.

account_tax_service_client

Generated client implementations.

accounts_service_client

Generated client implementations.

business_day_config

Nested message and enum types in BusinessDayConfig.

business_identity

Nested message and enum types in BusinessIdentity.

business_identity_service_client

Generated client implementations.

business_info_service_client

Generated client implementations.

create_and_configure_account_request

Nested message and enum types in CreateAndConfigureAccountRequest.

distance

Nested message and enum types in Distance.

email_preferences

Nested message and enum types in EmailPreferences.

email_preferences_service_client

Generated client implementations.

homepage_service_client

Generated client implementations.

minimum_order_value_table

Nested message and enum types in MinimumOrderValueTable.

online_return_policy

Nested message and enum types in OnlineReturnPolicy.

online_return_policy_service_client

Generated client implementations.

program

Nested message and enum types in Program.

programs_service_client

Generated client implementations.

region

Nested message and enum types in Region.

regions_service_client

Generated client implementations.

service

Nested message and enum types in Service.

shipping_settings_service_client

Generated client implementations.

tax_rule

Nested message and enum types in TaxRule.

terms_of_service_agreement_state_service_client

Generated client implementations.

terms_of_service_service_client

Generated client implementations.

transit_table

Nested message and enum types in TransitTable.

user

Nested message and enum types in User.

user_service_client

Generated client implementations.

Structs

AcceptTermsOfServiceRequest

Request message for the AcceptTermsOfService method.

Accepted

Describes the accepted terms of service.

Account

An account.

AccountIssue

An AccountIssue.

AccountTax

The tax settings of a merchant account. All methods require the admin role.

Address

Shipping address of the warehouse.

BusinessDayConfig

Business days of the warehouse.

BusinessIdentity

Collection of information related to the identity of a business.

BusinessInfo

Collection of information related to a business.

CarrierRate

A list of carrier rates that can be referred to by main_table or single_value.

ClaimHomepageRequest

Request message for the ClaimHomepage method.

CreateAndConfigureAccountRequest

Request message for the CreateAndConfigureAccount method.

CreateRegionRequest

Request message for the CreateRegion method.

CreateUserRequest

Request message for the CreateUser method.

CustomerService

Customer service information.

CutoffTime

Business days cutoff time definition.

DeleteAccountRequest

Request message for the DeleteAccount method.

DeleteRegionRequest

Request message for the DeleteRegion method.

DeleteUserRequest

Request message for the DeleteUser method.

DeliveryTime

Time spent in various aspects from order to the delivery of the product.

DisableProgramRequest

Request message for the DisableProgram method.

Distance

Maximum delivery radius. This is only required for the local delivery shipment type.

EmailPreferences

The categories of notifications the user opted into / opted out of. The email preferences do not include mandatory announcements as users can’t opt out of them.

EnableProgramRequest

Request message for the EnableProgram method.

GetAccountRequest

Request message for the GetAccount method.

GetAccountTaxRequest

Request to get tax settings

GetBusinessIdentityRequest

Request message for the GetBusinessIdentity method.

GetBusinessInfoRequest

Request message for the GetBusinessInfo method.

GetEmailPreferencesRequest

Request message for GetEmailPreferences method.

GetHomepageRequest

Request message for the GetHomepage method.

GetOnlineReturnPolicyRequest

Request message for the GetOnlineReturnPolicy method.

GetProgramRequest

Request message for the GetProgram method.

GetRegionRequest

Request message for the GetRegion method.

GetShippingSettingsRequest

Request message for the GetShippingSetting method.

GetTermsOfServiceAgreementStateRequest

Request message for the GetTermsOfServiceAgreementState method.

GetTermsOfServiceRequest

Request message for the GetTermsOfService method.

GetUserRequest

Request message for the GetUser method.

Headers

A non-empty list of row or column headers for a table. Exactly one of prices, weights, num_items, postal_code_group_names, or location must be set.

Homepage

A store’s homepage.

InsertShippingSettingsRequest

Request message for the InsertShippingSetting method.

ListAccountIssuesRequest

Request message for the ListAccountIssues method.

ListAccountIssuesResponse

Response message for the ListAccountIssues method.

ListAccountTaxRequest

Request to list all sub-account tax settings only for the requesting merchant This method can only be called on a multi-client account, otherwise it’ll return an error.

ListAccountTaxResponse

Response to account tax list request This method can only be called on a multi-client account, otherwise it’ll return an error.

ListAccountsRequest

Request message for the ListAccounts method.

ListAccountsResponse

Response message for the ListAccounts method.

ListOnlineReturnPoliciesRequest

Request message for the ListOnlineReturnPolicies method.

ListOnlineReturnPoliciesResponse

Response message for the ListOnlineReturnPolicies method.

ListProgramsRequest

Request message for the ListPrograms method.

ListProgramsResponse

Response message for the ListPrograms method.

ListRegionsRequest

Request message for the ListRegions method.

ListRegionsResponse

Response message for the ListRegions method.

ListSubAccountsRequest

Request message for the ListSubAccounts method.

ListSubAccountsResponse

Response message for the ListSubAccounts method.

ListUsersRequest

Request message for the ListUsers method.

ListUsersResponse

Response message for the ListUsers method.

LocationIdSet

A list of location ID sets. Must be non-empty. Can only be set if all other fields are not set.

MinimumOrderValueTable

Table of per store minimum order values for the pickup fulfillment type.

OnlineReturnPolicy

Online return policy object. This is currently used to represent return policies for ads and free listings programs.

Program

Defines participation in a given program for the specified account.

RateGroup

Shipping rate group definitions. Only the last one is allowed to have an empty applicable_shipping_labels, which means “everything else”. The other applicable_shipping_labels must not overlap.

Region

Represents a geographic region that you can use as a target with both the RegionalInventory and ShippingSettings services. You can define regions as collections of either postal codes or, in some countries, using predefined geotargets. For more information, see Set up regions for more information.

Required

Describes the terms of service which are required to be accepted.

RetrieveForApplicationTermsOfServiceAgreementStateRequest

Request message for the RetrieveForApplicationTermsOfServiceAgreementState method.

RetrieveLatestTermsOfServiceRequest

Request message for the RetrieveLatestTermsOfService method.

Row

Include a list of cells.

Service

Shipping service.

ShippingSettings

The merchant account’s [shipping setting]((https://support.google.com/merchants/answer/6069284).

Table

A table defining the rate group, when single_value is not expressive enough.

TaxRule

Primary type convension

TermsOfService

A TermsOfService.

TermsOfServiceAgreementState

This resource represents the agreement state for a given account and terms of service kind. The state is as follows:

TransitTable

Transit time table, number of business days spent in transit based on row and column dimensions. Either min_transit_days, max_transit_days or transit_time_table can be set, but not both.

UnclaimHomepageRequest

Request message for the UnclaimHomepage method.

UpdateAccountRequest

Request message for the UpdateAccount method.

UpdateAccountTaxRequest

Request to update the tax settings

UpdateBusinessIdentityRequest

Request message for the UpdateBusinessIdentity method.

UpdateBusinessInfoRequest

Request message for the UpdateBusinessInfo method.

UpdateEmailPreferencesRequest

Request message for UpdateEmailPreferences method.

UpdateHomepageRequest

Request message for the UpdateHomepage method.

UpdateRegionRequest

Request message for the UpdateRegion method.

UpdateUserRequest

Request message for the UpdateUser method.

User

A user.

Value

The single value of a rate group or the value of a rate group table’s cell. Exactly one of no_shipping, flat_rate, price_percentage, carrier_rateName, subtable_name must be set.

Warehouse

A fulfillment warehouse, which stores and handles inventory. Next tag: 7

WarehouseBasedDeliveryTime

Indicates that the delivery time should be calculated per warehouse (shipping origin location) based on the settings of the selected carrier. When set, no other transit time related field in delivery_time should be set.

WarehouseCutoffTime

The latest time of day that an order can be accepted and begin processing. Later orders will be processed in the next day. The time is based on the warehouse postal code.

Enums

AccessRight

The access right.

PhoneVerificationState

The phone verification state.

TermsOfServiceKind

The TermsOfService Kind.