Module google_api_proto::google::shopping::merchant::conversions::v1beta

Modules

attribution_settings

Nested message and enum types in AttributionSettings.

conversion_source

Nested message and enum types in ConversionSource.

conversion_sources_service_client

Generated client implementations.

Structs

AttributionSettings

Represents attribution settings for conversion sources receiving pre-attribution data.

ConversionSource

Represents a conversion source owned by a Merchant account. A merchant account can have up to 200 conversion sources.

CreateConversionSourceRequest

Request message for the CreateConversionSource method.

DeleteConversionSourceRequest

Request message for the DeleteConversionSource method.

GetConversionSourceRequest

Request message for the GetConversionSource method.

GoogleAnalyticsLink

“Google Analytics Link” sources can be used to get conversion data from an existing Google Analytics property into the linked Merchant Center account.

ListConversionSourcesRequest

Request message for the ListConversionSources method.

ListConversionSourcesResponse

Response message for the ListConversionSources method.

MerchantCenterDestination

“Merchant Center Destination” sources can be used to send conversion events from an online store using a Google tag directly to a Merchant Center account where the source is created.

UndeleteConversionSourceRequest

Request message for the UndeleteConversionSource method.

UpdateConversionSourceRequest

Request message for the UpdateConversionSource method.