Module google_api_proto::google::shopping::merchant::datasources::v1beta

Modules

data_source

Nested message and enum types in DataSource.

data_sources_service_client

Generated client implementations.

file_input

Nested message and enum types in FileInput.

primary_product_data_source

Nested message and enum types in PrimaryProductDataSource.

Structs

CreateDataSourceRequest

Request message for the CreateDataSource method.

DataSource

The data source for the Merchant Center account.

DeleteDataSourceRequest

Request message for the DeleteDataSource method.

FetchDataSourceRequest

Request message for the FetchDataSource method.

FileInput

The data specific for file data sources. This field is empty for other data source inputs.

GetDataSourceRequest

Request message for the GetDataSource method.

ListDataSourcesRequest

Request message for the ListDataSources method.

ListDataSourcesResponse

Response message for the ListDataSources method.

LocalInventoryDataSource

The local inventory data source.

PrimaryProductDataSource

The primary data source for local and online products.

PromotionDataSource

The promotion data source.

RegionalInventoryDataSource

The regional inventory data source.

SupplementalProductDataSource

The supplemental data source for local and online products.

UpdateDataSourceRequest

Request message for the UpdateDataSource method.