Module google_api_proto::google::shopping::merchant::products::v1beta

Modules

product_inputs_service_client

Generated client implementations.

product_status

Nested message and enum types in ProductStatus.

products_service_client

Generated client implementations.

Structs

Attributes

Attributes.

Certification

Product certification, initially introduced for EU energy efficiency labeling compliance using the EU EPREL database.

CloudExportAdditionalProperties

Product property for the Cloud Retail API. For example, properties for a TV product could be “Screen-Resolution” or “Screen-Size”.

DeleteProductInputRequest

Request message for the DeleteProductInput method.

FreeShippingThreshold

Conditions to be met for a product to have free shipping.

GetProductRequest

Request message for the GetProduct method.

InsertProductInputRequest

Request message for the InsertProductInput method.

Installment

A message that represents installment.

ListProductsRequest

Request message for the ListProducts method.

ListProductsResponse

Response message for the ListProducts method.

LoyaltyPoints

A message that represents loyalty points.

LoyaltyProgram

A message that represents loyalty program.

Product

The processed product, built from multiple [product inputs][[google.shopping.content.bundles.Products.ProductInput] after applying rules and supplemental data sources. This processed product matches what is shown in your Merchant Center account and in Shopping ads and other surfaces across Google. Each product is built from exactly one primary data source product input, and multiple supplemental data source inputs. After inserting, updating, or deleting a product input, it may take several minutes before the updated processed product can be retrieved.

ProductDetail

The product details.

ProductDimension

The dimension of the product.

ProductInput

This resource represents input data you submit for a product, not the processed product that you see in Merchant Center, in Shopping ads, or across Google surfaces. Product inputs, rules and supplemental data source data are combined to create the processed [product][google.shopping.content.bundles.Products.Product].

ProductStatus

The status of a product, data validation issues, that is, information about a product computed asynchronously.

ProductStructuredDescription

Structured description, for algorithmically (AI)-generated descriptions.

ProductStructuredTitle

Structured title, for algorithmically (AI)-generated titles.

ProductWeight

The weight of the product.

Shipping

The Shipping of the product.

ShippingDimension

The ShippingDimension of the product.

ShippingWeight

The ShippingWeight of the product.

SubscriptionCost

The SubscriptionCost of the product.

Tax

The Tax of the product.

UnitPricingBaseMeasure

The UnitPricingBaseMeasure of the product.

UnitPricingMeasure

The UnitPricingMeasure of the product.

Enums

SubscriptionPeriod

The subscription period of the product.