Struct google_api_proto::google::shopping::merchant::products::v1beta::Attributes

pub struct Attributes {

    pub identifier_exists: Option<bool>,
    pub is_bundle: Option<bool>,
    pub title: Option<String>,
    pub description: Option<String>,
    pub link: Option<String>,
    pub mobile_link: Option<String>,
    pub canonical_link: Option<String>,
    pub image_link: Option<String>,
    pub additional_image_links: Vec<String>,
    pub expiration_date: Option<Timestamp>,
    pub disclosure_date: Option<Timestamp>,
    pub adult: Option<bool>,
    pub age_group: Option<String>,
    pub availability: Option<String>,
    pub availability_date: Option<Timestamp>,
    pub brand: Option<String>,
    pub color: Option<String>,
    pub condition: Option<String>,
    pub gender: Option<String>,
    pub google_product_category: Option<String>,
    pub gtin: Option<String>,
    pub item_group_id: Option<String>,
    pub material: Option<String>,
    pub mpn: Option<String>,
    pub pattern: Option<String>,
    pub price: Option<Price>,
    pub installment: Option<Installment>,
    pub subscription_cost: Option<SubscriptionCost>,
    pub loyalty_points: Option<LoyaltyPoints>,
    pub loyalty_programs: Vec<LoyaltyProgram>,
    pub product_types: Vec<String>,
    pub sale_price: Option<Price>,
    pub sale_price_effective_date: Option<Interval>,
    pub sell_on_google_quantity: Option<i64>,
    pub product_height: Option<ProductDimension>,
    pub product_length: Option<ProductDimension>,
    pub product_width: Option<ProductDimension>,
    pub product_weight: Option<ProductWeight>,
    pub shipping: Vec<Shipping>,
    pub free_shipping_threshold: Vec<FreeShippingThreshold>,
    pub shipping_weight: Option<ShippingWeight>,
    pub shipping_length: Option<ShippingDimension>,
    pub shipping_width: Option<ShippingDimension>,
    pub shipping_height: Option<ShippingDimension>,
    pub max_handling_time: Option<i64>,
    pub min_handling_time: Option<i64>,
    pub shipping_label: Option<String>,
    pub transit_time_label: Option<String>,
    pub size: Option<String>,
    pub size_system: Option<String>,
    pub size_types: Vec<String>,
    pub taxes: Vec<Tax>,
    pub tax_category: Option<String>,
    pub energy_efficiency_class: Option<String>,
    pub min_energy_efficiency_class: Option<String>,
    pub max_energy_efficiency_class: Option<String>,
    pub unit_pricing_measure: Option<UnitPricingMeasure>,
    pub unit_pricing_base_measure: Option<UnitPricingBaseMeasure>,
    pub multipack: Option<i64>,
    pub ads_grouping: Option<String>,
    pub ads_labels: Vec<String>,
    pub ads_redirect: Option<String>,
    pub cost_of_goods_sold: Option<Price>,
    pub product_details: Vec<ProductDetail>,
    pub product_highlights: Vec<String>,
    pub display_ads_id: Option<String>,
    pub display_ads_similar_ids: Vec<String>,
    pub display_ads_title: Option<String>,
    pub display_ads_link: Option<String>,
    pub display_ads_value: Option<f64>,
    pub promotion_ids: Vec<String>,
    pub pickup_method: Option<String>,
    pub pickup_sla: Option<String>,
    pub link_template: Option<String>,
    pub mobile_link_template: Option<String>,
    pub custom_label_0: Option<String>,
    pub custom_label_1: Option<String>,
    pub custom_label_2: Option<String>,
    pub custom_label_3: Option<String>,
    pub custom_label_4: Option<String>,
    pub included_destinations: Vec<String>,
    pub excluded_destinations: Vec<String>,
    pub shopping_ads_excluded_countries: Vec<String>,
    pub external_seller_id: Option<String>,
    pub pause: Option<String>,
    pub lifestyle_image_links: Vec<String>,
    pub cloud_export_additional_properties: Vec<CloudExportAdditionalProperties>,
    pub virtual_model_link: Option<String>,
    pub certifications: Vec<Certification>,
    pub structured_title: Option<ProductStructuredTitle>,
    pub structured_description: Option<ProductStructuredDescription>,
    pub auto_pricing_min_price: Option<Price>,

}

Attributes.

Fields
identifier_exists: Option<bool>

Set this value to false when the item does not have unique product identifiers appropriate to its category, such as GTIN, MPN, and brand. Defaults to true, if not provided.

is_bundle: Option<bool>

Whether the item is a merchant-defined bundle. A bundle is a custom grouping of different products sold by a merchant for a single price.

title: Option<String>

Title of the item.

description: Option<String>

Description of the item.

link: Option<String>

URL directly linking to your item’s page on your online store.

mobile_link: Option<String>

URL for the mobile-optimized version of your item’s landing page.

canonical_link: Option<String>

URL for the canonical version of your item’s landing page.

image_link: Option<String>

URL of an image of the item.

additional_image_links: Vec<String>

Additional URLs of images of the item.

expiration_date: Option<Timestamp>

Date on which the item should expire, as specified upon insertion, in ISO 8601 format. The actual expiration date is exposed in productstatuses as googleExpirationDate and might be earlier if expirationDate is too far in the future.

disclosure_date: Option<Timestamp>

The date time when an offer becomes visible in search results across Google’s YouTube surfaces, in ISO 8601 format. See Disclosure date for more information.

adult: Option<bool>

Set to true if the item is targeted towards adults.

age_group: Option<String>

Target age group of the item.

availability: Option<String>

Availability status of the item.

availability_date: Option<Timestamp>

The day a pre-ordered product becomes available for delivery, in ISO 8601 format.

brand: Option<String>

Brand of the item.

color: Option<String>

Color of the item.

condition: Option<String>

Condition or state of the item.

gender: Option<String>

Target gender of the item.

google_product_category: Option<String>

Google’s category of the item (see Google product taxonomy). When querying products, this field will contain the user provided value. There is currently no way to get back the auto assigned google product categories through the API.

gtin: Option<String>

Global Trade Item Number (GTIN) of the item.

item_group_id: Option<String>

Shared identifier for all variants of the same product.

material: Option<String>

The material of which the item is made.

mpn: Option<String>

Manufacturer Part Number (MPN) of the item.

pattern: Option<String>

The item’s pattern (for example, polka dots).

price: Option<Price>

Price of the item.

installment: Option<Installment>

Number and amount of installments to pay for an item.

subscription_cost: Option<SubscriptionCost>

Number of periods (months or years) and amount of payment per period for an item with an associated subscription contract.

loyalty_points: Option<LoyaltyPoints>

Loyalty points that users receive after purchasing the item. Japan only.

loyalty_programs: Vec<LoyaltyProgram>

A list of loyalty program information that is used to surface loyalty benefits (for example, better pricing, points, etc) to the user of this item.

product_types: Vec<String>

Categories of the item (formatted as in product data specification).

sale_price: Option<Price>

Advertised sale price of the item.

sale_price_effective_date: Option<Interval>

Date range during which the item is on sale (see product data specification).

sell_on_google_quantity: Option<i64>

The quantity of the product that is available for selling on Google. Supported only for online products.

product_height: Option<ProductDimension>

The height of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_length: Option<ProductDimension>

The length of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_width: Option<ProductDimension>

The width of the product in the units provided. The value must be between 0 (exclusive) and 3000 (inclusive).

product_weight: Option<ProductWeight>

The weight of the product in the units provided. The value must be between 0 (exclusive) and 2000 (inclusive).

shipping: Vec<Shipping>

Shipping rules.

free_shipping_threshold: Vec<FreeShippingThreshold>

Conditions to be met for a product to have free shipping.

shipping_weight: Option<ShippingWeight>

Weight of the item for shipping.

shipping_length: Option<ShippingDimension>

Length of the item for shipping.

shipping_width: Option<ShippingDimension>

Width of the item for shipping.

shipping_height: Option<ShippingDimension>

Height of the item for shipping.

max_handling_time: Option<i64>

Maximal product handling time (in business days).

min_handling_time: Option<i64>

Minimal product handling time (in business days).

shipping_label: Option<String>

The shipping label of the product, used to group product in account-level shipping rules.

transit_time_label: Option<String>

The transit time label of the product, used to group product in account-level transit time tables.

size: Option<String>

Size of the item. Only one value is allowed. For variants with different sizes, insert a separate product for each size with the same itemGroupId value (see [https://support.google.com/merchants/answer/6324492](size definition)).

size_system: Option<String>

System in which the size is specified. Recommended for apparel items.

size_types: Vec<String>

The cut of the item. It can be used to represent combined size types for apparel items. Maximum two of size types can be provided (see [https://support.google.com/merchants/answer/6324497](size type)).

taxes: Vec<Tax>

Tax information.

tax_category: Option<String>

The tax category of the product, used to configure detailed tax nexus in account-level tax settings.

energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

min_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

max_energy_efficiency_class: Option<String>

The energy efficiency class as defined in EU directive 2010/30/EU.

unit_pricing_measure: Option<UnitPricingMeasure>

The measure and dimension of an item.

unit_pricing_base_measure: Option<UnitPricingBaseMeasure>

The preference of the denominator of the unit price.

multipack: Option<i64>

The number of identical products in a merchant-defined multipack.

ads_grouping: Option<String>

Used to group items in an arbitrary way. Only for CPA%, discouraged otherwise.

ads_labels: Vec<String>

Similar to ads_grouping, but only works on CPC.

ads_redirect: Option<String>

Allows advertisers to override the item URL when the product is shown within the context of Product ads.

cost_of_goods_sold: Option<Price>

Cost of goods sold. Used for gross profit reporting.

product_details: Vec<ProductDetail>

Technical specification or additional product details.

product_highlights: Vec<String>

Bullet points describing the most relevant highlights of a product.

display_ads_id: Option<String>

An identifier for an item for dynamic remarketing campaigns.

display_ads_similar_ids: Vec<String>

Advertiser-specified recommendations.

display_ads_title: Option<String>

Title of an item for dynamic remarketing campaigns.

display_ads_link: Option<String>

URL directly to your item’s landing page for dynamic remarketing campaigns.

display_ads_value: Option<f64>

Offer margin for dynamic remarketing campaigns.

promotion_ids: Vec<String>

The unique ID of a promotion.

pickup_method: Option<String>

The pick up option for the item.

pickup_sla: Option<String>

Item store pickup timeline.

link_template: Option<String>

Link template for merchant hosted local storefront.

mobile_link_template: Option<String>

Link template for merchant hosted local storefront optimized for mobile devices.

custom_label_0: Option<String>

Custom label 0 for custom grouping of items in a Shopping campaign.

custom_label_1: Option<String>

Custom label 1 for custom grouping of items in a Shopping campaign.

custom_label_2: Option<String>

Custom label 2 for custom grouping of items in a Shopping campaign.

custom_label_3: Option<String>

Custom label 3 for custom grouping of items in a Shopping campaign.

custom_label_4: Option<String>

Custom label 4 for custom grouping of items in a Shopping campaign.

included_destinations: Vec<String>

The list of destinations to include for this target (corresponds to checked check boxes in Merchant Center). Default destinations are always included unless provided in excludedDestinations.

excluded_destinations: Vec<String>

The list of destinations to exclude for this target (corresponds to unchecked check boxes in Merchant Center).

shopping_ads_excluded_countries: Vec<String>

List of country codes (ISO 3166-1 alpha-2) to exclude the offer from Shopping Ads destination. Countries from this list are removed from countries configured in data source settings.

external_seller_id: Option<String>

Required for multi-seller accounts. Use this attribute if you’re a marketplace uploading products for various sellers to your multi-seller account.

pause: Option<String>

Publication of this item will be temporarily paused.

lifestyle_image_links: Vec<String>

Additional URLs of lifestyle images of the item, used to explicitly identify images that showcase your item in a real-world context. See the Help Center article for more information.

cloud_export_additional_properties: Vec<CloudExportAdditionalProperties>

Extra fields to export to the Cloud Retail program.

virtual_model_link: Option<String>

URL of the 3D image of the item. See the Help Center article for more information.

certifications: Vec<Certification>

Product Certifications, for example for energy efficiency labeling of products recorded in the EU EPREL database. See the Help Center article for more information.

structured_title: Option<ProductStructuredTitle>

Structured title, for algorithmically (AI)-generated titles.

structured_description: Option<ProductStructuredDescription>

Structured description, for algorithmically (AI)-generated descriptions.

auto_pricing_min_price: Option<Price>

A safeguard in the “Automated Discounts” (https://support.google.com/merchants/answer/10295759) and “Dynamic Promotions” (https://support.google.com/merchants/answer/13949249) projects, ensuring that discounts on merchants’ offers do not fall below this value, thereby preserving the offer’s value and profitability.

Implementations

impl Attributes

pub fn external_seller_id(&self) -> &str

Returns the value of external_seller_id, or the default value if external_seller_id is unset.

pub fn identifier_exists(&self) -> bool

Returns the value of identifier_exists, or the default value if identifier_exists is unset.

pub fn is_bundle(&self) -> bool

Returns the value of is_bundle, or the default value if is_bundle is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn link(&self) -> &str

Returns the value of link, or the default value if link is unset.

pub fn mobile_link(&self) -> &str

Returns the value of mobile_link, or the default value if mobile_link is unset.

pub fn canonical_link(&self) -> &str

Returns the value of canonical_link, or the default value if canonical_link is unset.

pub fn image_link(&self) -> &str

Returns the value of image_link, or the default value if image_link is unset.

pub fn pause(&self) -> &str

Returns the value of pause, or the default value if pause is unset.

pub fn adult(&self) -> bool

Returns the value of adult, or the default value if adult is unset.

pub fn age_group(&self) -> &str

Returns the value of age_group, or the default value if age_group is unset.

pub fn availability(&self) -> &str

Returns the value of availability, or the default value if availability is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn color(&self) -> &str

Returns the value of color, or the default value if color is unset.

pub fn condition(&self) -> &str

Returns the value of condition, or the default value if condition is unset.

pub fn gender(&self) -> &str

Returns the value of gender, or the default value if gender is unset.

pub fn google_product_category(&self) -> &str

Returns the value of google_product_category, or the default value if google_product_category is unset.

pub fn gtin(&self) -> &str

Returns the value of gtin, or the default value if gtin is unset.

pub fn item_group_id(&self) -> &str

Returns the value of item_group_id, or the default value if item_group_id is unset.

pub fn material(&self) -> &str

Returns the value of material, or the default value if material is unset.

pub fn mpn(&self) -> &str

Returns the value of mpn, or the default value if mpn is unset.

pub fn pattern(&self) -> &str

Returns the value of pattern, or the default value if pattern is unset.

pub fn sell_on_google_quantity(&self) -> i64

Returns the value of sell_on_google_quantity, or the default value if sell_on_google_quantity is unset.

pub fn max_handling_time(&self) -> i64

Returns the value of max_handling_time, or the default value if max_handling_time is unset.

pub fn min_handling_time(&self) -> i64

Returns the value of min_handling_time, or the default value if min_handling_time is unset.

pub fn shipping_label(&self) -> &str

Returns the value of shipping_label, or the default value if shipping_label is unset.

pub fn transit_time_label(&self) -> &str

Returns the value of transit_time_label, or the default value if transit_time_label is unset.

pub fn size(&self) -> &str

Returns the value of size, or the default value if size is unset.

pub fn size_system(&self) -> &str

Returns the value of size_system, or the default value if size_system is unset.

pub fn tax_category(&self) -> &str

Returns the value of tax_category, or the default value if tax_category is unset.

pub fn energy_efficiency_class(&self) -> &str

Returns the value of energy_efficiency_class, or the default value if energy_efficiency_class is unset.

pub fn min_energy_efficiency_class(&self) -> &str

Returns the value of min_energy_efficiency_class, or the default value if min_energy_efficiency_class is unset.

pub fn max_energy_efficiency_class(&self) -> &str

Returns the value of max_energy_efficiency_class, or the default value if max_energy_efficiency_class is unset.

pub fn multipack(&self) -> i64

Returns the value of multipack, or the default value if multipack is unset.

pub fn ads_grouping(&self) -> &str

Returns the value of ads_grouping, or the default value if ads_grouping is unset.

pub fn ads_redirect(&self) -> &str

Returns the value of ads_redirect, or the default value if ads_redirect is unset.

pub fn display_ads_id(&self) -> &str

Returns the value of display_ads_id, or the default value if display_ads_id is unset.

pub fn display_ads_title(&self) -> &str

Returns the value of display_ads_title, or the default value if display_ads_title is unset.

pub fn display_ads_link(&self) -> &str

Returns the value of display_ads_link, or the default value if display_ads_link is unset.

pub fn display_ads_value(&self) -> f64

Returns the value of display_ads_value, or the default value if display_ads_value is unset.

pub fn custom_label_0(&self) -> &str

Returns the value of custom_label_0, or the default value if custom_label_0 is unset.

pub fn custom_label_1(&self) -> &str

Returns the value of custom_label_1, or the default value if custom_label_1 is unset.

pub fn custom_label_2(&self) -> &str

Returns the value of custom_label_2, or the default value if custom_label_2 is unset.

pub fn custom_label_3(&self) -> &str

Returns the value of custom_label_3, or the default value if custom_label_3 is unset.

pub fn custom_label_4(&self) -> &str

Returns the value of custom_label_4, or the default value if custom_label_4 is unset.

pub fn pickup_method(&self) -> &str

Returns the value of pickup_method, or the default value if pickup_method is unset.

pub fn pickup_sla(&self) -> &str

Returns the value of pickup_sla, or the default value if pickup_sla is unset.

pub fn link_template(&self) -> &str

Returns the value of link_template, or the default value if link_template is unset.

pub fn mobile_link_template(&self) -> &str

Returns the value of mobile_link_template, or the default value if mobile_link_template is unset.

pub fn virtual_model_link(&self) -> &str

Returns the value of virtual_model_link, or the default value if virtual_model_link is unset.