Struct google_api_proto::google::shopping::merchant::reports::v1beta::BestSellersBrandView

pub struct BestSellersBrandView {
    pub report_date: Option<Date>,
    pub report_granularity: Option<i32>,
    pub report_country_code: Option<String>,
    pub report_category_id: Option<i64>,
    pub brand: Option<String>,
    pub rank: Option<i64>,
    pub previous_rank: Option<i64>,
    pub relative_demand: Option<i32>,
    pub previous_relative_demand: Option<i32>,
    pub relative_demand_change: Option<i32>,
}

Fields available for query in best_sellers_brand_view table.

Best sellers report with top brands.

Values are only set for fields requested explicitly in the request’s search query.

Fields
report_date: Option<Date>

Report date. The value of this field can only be one of the following:

The first day of the week (Monday) for weekly reports,

The first day of the month for monthly reports.

Required in the SELECT clause. If a WHERE condition on report_date is not specified in the query, the latest available weekly or monthly report is returned.

report_granularity: Option<i32>

Granularity of the report. The ranking can be done over a week or a month timeframe.

Required in the SELECT clause. Condition on report_granularity is required in the WHERE clause.

report_country_code: Option<String>

Country where the ranking is calculated. Represented in the ISO 3166 format.

Required in the SELECT clause. Condition on report_country_code is required in the WHERE clause.

report_category_id: Option<i64>

Google product category ID to calculate the ranking for, represented in Google’s product taxonomy.

Required in the SELECT clause. If a WHERE condition on report_category_id is not specified in the query, rankings for all top-level categories are returned.

brand: Option<String>

Name of the brand.

rank: Option<i64>

Popularity of the brand on Ads and organic surfaces, in the selected category and country, based on the estimated number of units sold.

previous_rank: Option<i64>

Popularity rank in the previous week or month.

relative_demand: Option<i32>

Estimated demand in relation to the brand with the highest popularity rank in the same category and country.

previous_relative_demand: Option<i32>

Estimated demand in relation to the brand with the highest popularity rank in the same category and country in the previous week or month.

relative_demand_change: Option<i32>

Change in the estimated demand. Whether it rose, sank or remained flat.

Implementations

impl BestSellersBrandView

pub fn report_granularity(&self) -> ReportGranularityEnum

Returns the enum value of report_granularity, or the default if the field is unset or set to an invalid enum value.

pub fn set_report_granularity(&mut self, value: ReportGranularityEnum)

Sets report_granularity to the provided enum value.

pub fn report_country_code(&self) -> &str

Returns the value of report_country_code, or the default value if report_country_code is unset.

pub fn report_category_id(&self) -> i64

Returns the value of report_category_id, or the default value if report_category_id is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn rank(&self) -> i64

Returns the value of rank, or the default value if rank is unset.

pub fn previous_rank(&self) -> i64

Returns the value of previous_rank, or the default value if previous_rank is unset.

pub fn relative_demand(&self) -> RelativeDemandEnum

Returns the enum value of relative_demand, or the default if the field is unset or set to an invalid enum value.

pub fn set_relative_demand(&mut self, value: RelativeDemandEnum)

Sets relative_demand to the provided enum value.

pub fn previous_relative_demand(&self) -> RelativeDemandEnum

Returns the enum value of previous_relative_demand, or the default if the field is unset or set to an invalid enum value.

pub fn set_previous_relative_demand(&mut self, value: RelativeDemandEnum)

Sets previous_relative_demand to the provided enum value.

pub fn relative_demand_change(&self) -> RelativeDemandChangeTypeEnum

Returns the enum value of relative_demand_change, or the default if the field is unset or set to an invalid enum value.

pub fn set_relative_demand_change( &mut self, value: RelativeDemandChangeTypeEnum, )

Sets relative_demand_change to the provided enum value.