Struct google_api_proto::google::shopping::merchant::reports::v1beta::PriceInsightsProductView

pub struct PriceInsightsProductView {

    pub id: Option<String>,
    pub offer_id: Option<String>,
    pub title: Option<String>,
    pub brand: Option<String>,
    pub category_l1: Option<String>,
    pub category_l2: Option<String>,
    pub category_l3: Option<String>,
    pub category_l4: Option<String>,
    pub category_l5: Option<String>,
    pub product_type_l1: Option<String>,
    pub product_type_l2: Option<String>,
    pub product_type_l3: Option<String>,
    pub product_type_l4: Option<String>,
    pub product_type_l5: Option<String>,
    pub price: Option<Price>,
    pub suggested_price: Option<Price>,
    pub predicted_impressions_change_fraction: Option<f64>,
    pub predicted_clicks_change_fraction: Option<f64>,
    pub predicted_conversions_change_fraction: Option<f64>,
    pub effectiveness: i32,

}

Fields available for query in price_insights_product_view table.

Price insights report.

Values are only set for fields requested explicitly in the request’s search query.

Fields
id: Option<String>

REST ID of the product, in the form of channel~languageCode~feedLabel~offerId. Can be used to join data with the product_view table.

Required in the SELECT clause.

offer_id: Option<String>

Merchant-provided id of the product.

title: Option<String>

Title of the product.

brand: Option<String>

Brand of the product.

category_l1: Option<String>

Product category (1st level) in Google’s product taxonomy.

category_l2: Option<String>

Product category (2nd level) in Google’s product taxonomy.

category_l3: Option<String>

Product category (3rd level) in Google’s product taxonomy.

category_l4: Option<String>

Product category (4th level) in Google’s product taxonomy.

category_l5: Option<String>

Product category (5th level) in Google’s product taxonomy.

product_type_l1: Option<String>

Product type (1st level) in merchant’s own product taxonomy.

product_type_l2: Option<String>

Product type (2nd level) in merchant’s own product taxonomy.

product_type_l3: Option<String>

Product type (3rd level) in merchant’s own product taxonomy.

product_type_l4: Option<String>

Product type (4th level) in merchant’s own product taxonomy.

product_type_l5: Option<String>

Product type (5th level) in merchant’s own product taxonomy.

price: Option<Price>

Current price of the product.

suggested_price: Option<Price>

Latest suggested price for the product.

predicted_impressions_change_fraction: Option<f64>

Predicted change in impressions as a fraction after introducing the suggested price compared to current active price. For example, 0.05 is a 5% predicted increase in impressions.

predicted_clicks_change_fraction: Option<f64>

Predicted change in clicks as a fraction after introducing the suggested price compared to current active price. For example, 0.05 is a 5% predicted increase in clicks.

predicted_conversions_change_fraction: Option<f64>

Predicted change in conversions as a fraction after introducing the suggested price compared to current active price. For example, 0.05 is a 5% predicted increase in conversions).

effectiveness: i32

The predicted effectiveness of applying the price suggestion, bucketed.

Implementations

impl PriceInsightsProductView

pub fn id(&self) -> &str

Returns the value of id, or the default value if id is unset.

pub fn offer_id(&self) -> &str

Returns the value of offer_id, or the default value if offer_id is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn category_l1(&self) -> &str

Returns the value of category_l1, or the default value if category_l1 is unset.

pub fn category_l2(&self) -> &str

Returns the value of category_l2, or the default value if category_l2 is unset.

pub fn category_l3(&self) -> &str

Returns the value of category_l3, or the default value if category_l3 is unset.

pub fn category_l4(&self) -> &str

Returns the value of category_l4, or the default value if category_l4 is unset.

pub fn category_l5(&self) -> &str

Returns the value of category_l5, or the default value if category_l5 is unset.

pub fn product_type_l1(&self) -> &str

Returns the value of product_type_l1, or the default value if product_type_l1 is unset.

pub fn product_type_l2(&self) -> &str

Returns the value of product_type_l2, or the default value if product_type_l2 is unset.

pub fn product_type_l3(&self) -> &str

Returns the value of product_type_l3, or the default value if product_type_l3 is unset.

pub fn product_type_l4(&self) -> &str

Returns the value of product_type_l4, or the default value if product_type_l4 is unset.

pub fn product_type_l5(&self) -> &str

Returns the value of product_type_l5, or the default value if product_type_l5 is unset.

pub fn predicted_impressions_change_fraction(&self) -> f64

Returns the value of predicted_impressions_change_fraction, or the default value if predicted_impressions_change_fraction is unset.

pub fn predicted_clicks_change_fraction(&self) -> f64

Returns the value of predicted_clicks_change_fraction, or the default value if predicted_clicks_change_fraction is unset.

pub fn predicted_conversions_change_fraction(&self) -> f64

Returns the value of predicted_conversions_change_fraction, or the default value if predicted_conversions_change_fraction is unset.

pub fn effectiveness(&self) -> Effectiveness

Returns the enum value of effectiveness, or the default if the field is set to an invalid enum value.

pub fn set_effectiveness(&mut self, value: Effectiveness)

Sets effectiveness to the provided enum value.