Struct google_api_proto::google::shopping::merchant::reports::v1beta::ProductPerformanceView

pub struct ProductPerformanceView {

    pub marketing_method: Option<i32>,
    pub date: Option<Date>,
    pub week: Option<Date>,
    pub customer_country_code: Option<String>,
    pub offer_id: Option<String>,
    pub title: Option<String>,
    pub brand: Option<String>,
    pub category_l1: Option<String>,
    pub category_l2: Option<String>,
    pub category_l3: Option<String>,
    pub category_l4: Option<String>,
    pub category_l5: Option<String>,
    pub product_type_l1: Option<String>,
    pub product_type_l2: Option<String>,
    pub product_type_l3: Option<String>,
    pub product_type_l4: Option<String>,
    pub product_type_l5: Option<String>,
    pub custom_label0: Option<String>,
    pub custom_label1: Option<String>,
    pub custom_label2: Option<String>,
    pub custom_label3: Option<String>,
    pub custom_label4: Option<String>,
    pub clicks: Option<i64>,
    pub impressions: Option<i64>,
    pub click_through_rate: Option<f64>,
    pub conversions: Option<f64>,
    pub conversion_value: Option<Price>,
    pub conversion_rate: Option<f64>,

}

Fields available for query in product_performance_view table.

Product performance data for your account, including performance metrics (for example, clicks) and dimensions according to which performance metrics are segmented (for example, offer_id). Values of product dimensions, such as offer_id, reflect the state of a product at the time of the impression.

Segment fields cannot be selected in queries without also selecting at least one metric field.

Values are only set for fields requested explicitly in the request’s search query.

Fields
marketing_method: Option<i32>

Marketing method to which metrics apply. Segment.

date: Option<Date>

Date in the merchant timezone to which metrics apply. Segment.

Condition on date is required in the WHERE clause.

week: Option<Date>

First day of the week (Monday) of the metrics date in the merchant timezone. Segment.

customer_country_code: Option<String>

Code of the country where the customer is located at the time of the event. Represented in the ISO 3166 format. Segment.

If the customer country cannot be determined, a special ‘ZZ’ code is returned.

offer_id: Option<String>

Merchant-provided id of the product. Segment.

title: Option<String>

Title of the product. Segment.

brand: Option<String>

Brand of the product. Segment.

category_l1: Option<String>

Product category (1st level) in Google’s product taxonomy. Segment.

category_l2: Option<String>

Product category (2nd level) in Google’s product taxonomy. Segment.

category_l3: Option<String>

Product category (3rd level) in Google’s product taxonomy. Segment.

category_l4: Option<String>

Product category (4th level) in Google’s product taxonomy. Segment.

category_l5: Option<String>

Product category (5th level) in Google’s product taxonomy. Segment.

product_type_l1: Option<String>

Product type (1st level) in merchant’s own product taxonomy. Segment.

product_type_l2: Option<String>

Product type (2nd level) in merchant’s own product taxonomy. Segment.

product_type_l3: Option<String>

Product type (3rd level) in merchant’s own product taxonomy. Segment.

product_type_l4: Option<String>

Product type (4th level) in merchant’s own product taxonomy. Segment.

product_type_l5: Option<String>

Product type (5th level) in merchant’s own product taxonomy. Segment.

custom_label0: Option<String>

Custom label 0 for custom grouping of products. Segment.

custom_label1: Option<String>

Custom label 1 for custom grouping of products. Segment.

custom_label2: Option<String>

Custom label 2 for custom grouping of products. Segment.

custom_label3: Option<String>

Custom label 3 for custom grouping of products. Segment.

custom_label4: Option<String>

Custom label 4 for custom grouping of products. Segment.

clicks: Option<i64>

Number of clicks. Metric.

impressions: Option<i64>

Number of times merchant’s products are shown. Metric.

click_through_rate: Option<f64>

Click-through rate - the number of clicks merchant’s products receive (clicks) divided by the number of times the products are shown (impressions). Metric.

conversions: Option<f64>

Number of conversions attributed to the product, reported on the conversion date. Depending on the attribution model, a conversion might be distributed across multiple clicks, where each click gets its own credit assigned. This metric is a sum of all such credits. Metric.

Available only for the FREE traffic source.

conversion_value: Option<Price>

Value of conversions attributed to the product, reported on the conversion date. Metric.

Available only for the FREE traffic source.

conversion_rate: Option<f64>

Number of conversions divided by the number of clicks, reported on the impression date. Metric.

Available only for the FREE traffic source.

Implementations

impl ProductPerformanceView

pub fn marketing_method(&self) -> MarketingMethodEnum

Returns the enum value of marketing_method, or the default if the field is unset or set to an invalid enum value.

pub fn set_marketing_method(&mut self, value: MarketingMethodEnum)

Sets marketing_method to the provided enum value.

pub fn customer_country_code(&self) -> &str

Returns the value of customer_country_code, or the default value if customer_country_code is unset.

pub fn offer_id(&self) -> &str

Returns the value of offer_id, or the default value if offer_id is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn brand(&self) -> &str

Returns the value of brand, or the default value if brand is unset.

pub fn category_l1(&self) -> &str

Returns the value of category_l1, or the default value if category_l1 is unset.

pub fn category_l2(&self) -> &str

Returns the value of category_l2, or the default value if category_l2 is unset.

pub fn category_l3(&self) -> &str

Returns the value of category_l3, or the default value if category_l3 is unset.

pub fn category_l4(&self) -> &str

Returns the value of category_l4, or the default value if category_l4 is unset.

pub fn category_l5(&self) -> &str

Returns the value of category_l5, or the default value if category_l5 is unset.

pub fn product_type_l1(&self) -> &str

Returns the value of product_type_l1, or the default value if product_type_l1 is unset.

pub fn product_type_l2(&self) -> &str

Returns the value of product_type_l2, or the default value if product_type_l2 is unset.

pub fn product_type_l3(&self) -> &str

Returns the value of product_type_l3, or the default value if product_type_l3 is unset.

pub fn product_type_l4(&self) -> &str

Returns the value of product_type_l4, or the default value if product_type_l4 is unset.

pub fn product_type_l5(&self) -> &str

Returns the value of product_type_l5, or the default value if product_type_l5 is unset.

pub fn custom_label0(&self) -> &str

Returns the value of custom_label0, or the default value if custom_label0 is unset.

pub fn custom_label1(&self) -> &str

Returns the value of custom_label1, or the default value if custom_label1 is unset.

pub fn custom_label2(&self) -> &str

Returns the value of custom_label2, or the default value if custom_label2 is unset.

pub fn custom_label3(&self) -> &str

Returns the value of custom_label3, or the default value if custom_label3 is unset.

pub fn custom_label4(&self) -> &str

Returns the value of custom_label4, or the default value if custom_label4 is unset.

pub fn clicks(&self) -> i64

Returns the value of clicks, or the default value if clicks is unset.

pub fn impressions(&self) -> i64

Returns the value of impressions, or the default value if impressions is unset.

pub fn click_through_rate(&self) -> f64

Returns the value of click_through_rate, or the default value if click_through_rate is unset.

pub fn conversions(&self) -> f64

Returns the value of conversions, or the default value if conversions is unset.

pub fn conversion_rate(&self) -> f64

Returns the value of conversion_rate, or the default value if conversion_rate is unset.