Module google_api_proto::google::spanner::admin::database::v1

Modules

backup

Nested message and enum types in Backup.

backup_schedule

Nested message and enum types in BackupSchedule.

backup_schedule_spec

Nested message and enum types in BackupScheduleSpec.

copy_backup_encryption_config

Nested message and enum types in CopyBackupEncryptionConfig.

create_backup_encryption_config

Nested message and enum types in CreateBackupEncryptionConfig.

database

Nested message and enum types in Database.

database_admin_client

Generated client implementations.

encryption_info

Nested message and enum types in EncryptionInfo.

restore_database_encryption_config

Nested message and enum types in RestoreDatabaseEncryptionConfig.

restore_database_metadata

Nested message and enum types in RestoreDatabaseMetadata.

restore_database_request

Nested message and enum types in RestoreDatabaseRequest.

restore_info

Nested message and enum types in RestoreInfo.

Structs

Backup

A backup of a Cloud Spanner database.

BackupInfo

Information about a backup.

BackupSchedule

BackupSchedule expresses the automated backup creation specification for a Spanner database. Next ID: 10

BackupScheduleSpec

Defines specifications of the backup schedule.

CopyBackupEncryptionConfig

Encryption configuration for the copied backup.

CopyBackupMetadata

Metadata type for the operation returned by [CopyBackup][google.spanner.admin.database.v1.DatabaseAdmin.CopyBackup].

CopyBackupRequest

The request for [CopyBackup][google.spanner.admin.database.v1.DatabaseAdmin.CopyBackup].

CreateBackupEncryptionConfig

Encryption configuration for the backup to create.

CreateBackupMetadata

Metadata type for the operation returned by [CreateBackup][google.spanner.admin.database.v1.DatabaseAdmin.CreateBackup].

CreateBackupRequest

The request for [CreateBackup][google.spanner.admin.database.v1.DatabaseAdmin.CreateBackup].

CreateBackupScheduleRequest

The request for [CreateBackupSchedule][google.spanner.admin.database.v1.DatabaseAdmin.CreateBackupSchedule].

CreateDatabaseMetadata

Metadata type for the operation returned by [CreateDatabase][google.spanner.admin.database.v1.DatabaseAdmin.CreateDatabase].

CreateDatabaseRequest

The request for [CreateDatabase][google.spanner.admin.database.v1.DatabaseAdmin.CreateDatabase].

CrontabSpec

CrontabSpec can be used to specify the version time and frequency at which the backup should be created.

Database

A Cloud Spanner database.

DatabaseRole

A Cloud Spanner database role.

DdlStatementActionInfo

Action information extracted from a DDL statement. This proto is used to display the brief info of the DDL statement for the operation [UpdateDatabaseDdl][google.spanner.admin.database.v1.DatabaseAdmin.UpdateDatabaseDdl].

DeleteBackupRequest

The request for [DeleteBackup][google.spanner.admin.database.v1.DatabaseAdmin.DeleteBackup].

DeleteBackupScheduleRequest

The request for [DeleteBackupSchedule][google.spanner.admin.database.v1.DatabaseAdmin.DeleteBackupSchedule].

DropDatabaseRequest

The request for [DropDatabase][google.spanner.admin.database.v1.DatabaseAdmin.DropDatabase].

EncryptionConfig

Encryption configuration for a Cloud Spanner database.

EncryptionInfo

Encryption information for a Cloud Spanner database or backup.

FullBackupSpec

The specification for full backups. A full backup stores the entire contents of the database at a given version time.

GetBackupRequest

The request for [GetBackup][google.spanner.admin.database.v1.DatabaseAdmin.GetBackup].

GetBackupScheduleRequest

The request for [GetBackupSchedule][google.spanner.admin.database.v1.DatabaseAdmin.GetBackupSchedule].

GetDatabaseDdlRequest

The request for [GetDatabaseDdl][google.spanner.admin.database.v1.DatabaseAdmin.GetDatabaseDdl].

GetDatabaseDdlResponse

The response for [GetDatabaseDdl][google.spanner.admin.database.v1.DatabaseAdmin.GetDatabaseDdl].

GetDatabaseRequest

The request for [GetDatabase][google.spanner.admin.database.v1.DatabaseAdmin.GetDatabase].

IncrementalBackupSpec

The specification for incremental backup chains. An incremental backup stores the delta of changes between a previous backup and the database contents at a given version time. An incremental backup chain consists of a full backup and zero or more successive incremental backups. The first backup created for an incremental backup chain is always a full backup.

ListBackupOperationsRequest

The request for [ListBackupOperations][google.spanner.admin.database.v1.DatabaseAdmin.ListBackupOperations].

ListBackupOperationsResponse

The response for [ListBackupOperations][google.spanner.admin.database.v1.DatabaseAdmin.ListBackupOperations].

ListBackupSchedulesRequest

The request for [ListBackupSchedules][google.spanner.admin.database.v1.DatabaseAdmin.ListBackupSchedules].

ListBackupSchedulesResponse

The response for [ListBackupSchedules][google.spanner.admin.database.v1.DatabaseAdmin.ListBackupSchedules].

ListBackupsRequest

The request for [ListBackups][google.spanner.admin.database.v1.DatabaseAdmin.ListBackups].

ListBackupsResponse

The response for [ListBackups][google.spanner.admin.database.v1.DatabaseAdmin.ListBackups].

ListDatabaseOperationsRequest

The request for [ListDatabaseOperations][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabaseOperations].

ListDatabaseOperationsResponse

The response for [ListDatabaseOperations][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabaseOperations].

ListDatabaseRolesRequest

The request for [ListDatabaseRoles][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabaseRoles].

ListDatabaseRolesResponse

The response for [ListDatabaseRoles][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabaseRoles].

ListDatabasesRequest

The request for [ListDatabases][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabases].

ListDatabasesResponse

The response for [ListDatabases][google.spanner.admin.database.v1.DatabaseAdmin.ListDatabases].

OperationProgress

Encapsulates progress related information for a Cloud Spanner long running operation.

OptimizeRestoredDatabaseMetadata

Metadata type for the long-running operation used to track the progress of optimizations performed on a newly restored database. This long-running operation is automatically created by the system after the successful completion of a database restore, and cannot be cancelled.

RestoreDatabaseEncryptionConfig

Encryption configuration for the restored database.

RestoreDatabaseMetadata

Metadata type for the long-running operation returned by [RestoreDatabase][google.spanner.admin.database.v1.DatabaseAdmin.RestoreDatabase].

RestoreDatabaseRequest

The request for [RestoreDatabase][google.spanner.admin.database.v1.DatabaseAdmin.RestoreDatabase].

RestoreInfo

Information about the database restore.

UpdateBackupRequest

The request for [UpdateBackup][google.spanner.admin.database.v1.DatabaseAdmin.UpdateBackup].

UpdateBackupScheduleRequest

The request for [UpdateBackupScheduleRequest][google.spanner.admin.database.v1.DatabaseAdmin.UpdateBackupSchedule].

UpdateDatabaseDdlMetadata

Metadata type for the operation returned by [UpdateDatabaseDdl][google.spanner.admin.database.v1.DatabaseAdmin.UpdateDatabaseDdl].

UpdateDatabaseDdlRequest

Enqueues the given DDL statements to be applied, in order but not necessarily all at once, to the database schema at some point (or points) in the future. The server checks that the statements are executable (syntactically valid, name tables that exist, etc.) before enqueueing them, but they may still fail upon later execution (e.g., if a statement from another batch of statements is applied first and it conflicts in some way, or if there is some data-related problem like a NULL value in a column to which NOT NULL would be added). If a statement fails, all subsequent statements in the batch are automatically cancelled.

UpdateDatabaseMetadata

Metadata type for the operation returned by [UpdateDatabase][google.spanner.admin.database.v1.DatabaseAdmin.UpdateDatabase].

UpdateDatabaseRequest

The request for [UpdateDatabase][google.spanner.admin.database.v1.DatabaseAdmin.UpdateDatabase].

Enums

DatabaseDialect

Indicates the dialect type of a database.

RestoreSourceType

Indicates the type of the restore source.