Module google_api_proto::google::spanner::admin::instance::v1

Modules

autoscaling_config

Nested message and enum types in AutoscalingConfig.

instance

Nested message and enum types in Instance.

instance_admin_client

Generated client implementations.

instance_config

Nested message and enum types in InstanceConfig.

instance_partition

Nested message and enum types in InstancePartition.

replica_info

Nested message and enum types in ReplicaInfo.

Structs

AutoscalingConfig

Autoscaling configuration for an instance.

CreateInstanceConfigMetadata

Metadata type for the operation returned by [CreateInstanceConfig][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstanceConfig].

CreateInstanceConfigRequest

The request for [CreateInstanceConfigRequest][InstanceAdmin.CreateInstanceConfigRequest].

CreateInstanceMetadata

Metadata type for the operation returned by [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

CreateInstancePartitionMetadata

Metadata type for the operation returned by [CreateInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstancePartition].

CreateInstancePartitionRequest

The request for [CreateInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstancePartition].

CreateInstanceRequest

The request for [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

DeleteInstanceConfigRequest

The request for [DeleteInstanceConfigRequest][InstanceAdmin.DeleteInstanceConfigRequest].

DeleteInstancePartitionRequest

The request for [DeleteInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstancePartition].

DeleteInstanceRequest

The request for [DeleteInstance][google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstance].

GetInstanceConfigRequest

The request for [GetInstanceConfigRequest][google.spanner.admin.instance.v1.InstanceAdmin.GetInstanceConfig].

GetInstancePartitionRequest

The request for [GetInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.GetInstancePartition].

GetInstanceRequest

The request for [GetInstance][google.spanner.admin.instance.v1.InstanceAdmin.GetInstance].

Instance

An isolated set of Cloud Spanner resources on which databases can be hosted.

InstanceConfig

A possible configuration for a Cloud Spanner instance. Configurations define the geographic placement of nodes and their replication.

InstancePartition

An isolated set of Cloud Spanner resources that databases can define placements on.

ListInstanceConfigOperationsRequest

The request for [ListInstanceConfigOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigOperations].

ListInstanceConfigOperationsResponse

The response for [ListInstanceConfigOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigOperations].

ListInstanceConfigsRequest

The request for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstanceConfigsResponse

The response for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstancePartitionOperationsRequest

The request for [ListInstancePartitionOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstancePartitionOperations].

ListInstancePartitionOperationsResponse

The response for [ListInstancePartitionOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstancePartitionOperations].

ListInstancePartitionsRequest

The request for [ListInstancePartitions][google.spanner.admin.instance.v1.InstanceAdmin.ListInstancePartitions].

ListInstancePartitionsResponse

The response for [ListInstancePartitions][google.spanner.admin.instance.v1.InstanceAdmin.ListInstancePartitions].

ListInstancesRequest

The request for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

ListInstancesResponse

The response for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

MoveInstanceMetadata

Metadata type for the operation returned by [MoveInstance][google.spanner.admin.instance.v1.InstanceAdmin.MoveInstance].

MoveInstanceRequest

The request for [MoveInstance][google.spanner.admin.instance.v1.InstanceAdmin.MoveInstance].

MoveInstanceResponse

The response for [MoveInstance][google.spanner.admin.instance.v1.InstanceAdmin.MoveInstance].

OperationProgress

Encapsulates progress related information for a Cloud Spanner long running instance operations.

ReplicaInfo

UpdateInstanceConfigMetadata

Metadata type for the operation returned by [UpdateInstanceConfig][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstanceConfig].

UpdateInstanceConfigRequest

The request for [UpdateInstanceConfigRequest][InstanceAdmin.UpdateInstanceConfigRequest].

UpdateInstanceMetadata

Metadata type for the operation returned by [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].

UpdateInstancePartitionMetadata

Metadata type for the operation returned by [UpdateInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstancePartition].

UpdateInstancePartitionRequest

The request for [UpdateInstancePartition][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstancePartition].

UpdateInstanceRequest

The request for [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].

Enums

FulfillmentPeriod

Indicates the expected fulfillment period of an operation.