Module google_api_proto::google::storage::control::v2

Modules

storage_control_client

Generated client implementations.

storage_layout

Nested message and enum types in StorageLayout.

Structs

CommonLongRunningOperationMetadata

The message contains metadata that is common to all Storage Control long-running operations, present in its google.longrunning.Operation messages, and accessible via metadata.common_metadata.

CreateFolderRequest

Request message for CreateFolder.

CreateManagedFolderRequest

Request message for CreateManagedFolder.

DeleteFolderRequest

Request message for DeleteFolder.

DeleteManagedFolderRequest

DeleteManagedFolder RPC request message.

Folder

A folder.

GetFolderRequest

Request message for GetFolder.

GetManagedFolderRequest

Request message for GetManagedFolder.

GetStorageLayoutRequest

Request message for GetStorageLayout.

ListFoldersRequest

Request message for ListFolders.

ListFoldersResponse

Response message for ListFolders.

ListManagedFoldersRequest

Request message for ListManagedFolders.

ListManagedFoldersResponse

Response message for ListManagedFolders.

ManagedFolder

A managed folder.

PendingRenameInfo

Contains information about a pending rename operation.

RenameFolderMetadata

Message returned in the metadata field of the Operation resource for RenameFolder operations.

RenameFolderRequest

Request message for RenameFolder.

StorageLayout

The storage layout configuration of a bucket.