Module google_api_proto::google::storage::v2

Modules

bidi_write_object_request

Nested message and enum types in BidiWriteObjectRequest.

bidi_write_object_response

Nested message and enum types in BidiWriteObjectResponse.

bucket

Nested message and enum types in Bucket.

compose_object_request

Nested message and enum types in ComposeObjectRequest.

query_write_status_response

Nested message and enum types in QueryWriteStatusResponse.

service_constants

Nested message and enum types in ServiceConstants.

storage_client

Generated client implementations.

write_object_request

Nested message and enum types in WriteObjectRequest.

write_object_response

Nested message and enum types in WriteObjectResponse.

Structs

BidiWriteObjectRequest

Request message for BidiWriteObject.

BidiWriteObjectResponse

Response message for BidiWriteObject.

Bucket

A bucket.

BucketAccessControl

An access-control entry.

CancelResumableWriteRequest

Message for canceling an in-progress resumable upload. upload_id must be set.

CancelResumableWriteResponse

Empty response message for canceling an in-progress resumable upload, will be extended as needed.

ChecksummedData

Message used to convey content being read or written, along with an optional checksum.

CommonObjectRequestParams

Parameters that can be passed to any object request.

ComposeObjectRequest

Request message for ComposeObject.

ContentRange

Specifies a requested range of bytes to download.

CreateBucketRequest

Request message for CreateBucket.

CreateHmacKeyRequest

Request message for CreateHmacKey.

CreateHmacKeyResponse

Create hmac response. The only time the secret for an HMAC will be returned.

CreateNotificationConfigRequest

Request message for CreateNotificationConfig.

CustomerEncryption

Describes the Customer-Supplied Encryption Key mechanism used to store an Object’s data at rest.

DeleteBucketRequest

Request message for DeleteBucket.

DeleteHmacKeyRequest

Request object to delete a given HMAC key.

DeleteNotificationConfigRequest

Request message for DeleteNotificationConfig.

DeleteObjectRequest

Message for deleting an object. bucket and object must be set.

GetBucketRequest

Request message for GetBucket.

GetHmacKeyRequest

Request object to get metadata on a given HMAC key.

GetNotificationConfigRequest

Request message for GetNotificationConfig.

GetObjectRequest

Request message for GetObject.

GetServiceAccountRequest

Request message for GetServiceAccount.

HmacKeyMetadata

Hmac Key Metadata, which includes all information other than the secret.

ListBucketsRequest

Request message for ListBuckets.

ListBucketsResponse

The result of a call to Buckets.ListBuckets

ListHmacKeysRequest

Request to fetch a list of HMAC keys under a given project.

ListHmacKeysResponse

Hmac key list response with next page information.

ListNotificationConfigsRequest

Request message for ListNotifications.

ListNotificationConfigsResponse

The result of a call to ListNotificationConfigs

ListObjectsRequest

Request message for ListObjects.

ListObjectsResponse

The result of a call to Objects.ListObjects

LockBucketRetentionPolicyRequest

Request message for LockBucketRetentionPolicyRequest.

NotificationConfig

A directive to publish Pub/Sub notifications upon changes to a bucket.

Object

An object.

ObjectAccessControl

An access-control entry.

ObjectChecksums

Message used for storing full (not subrange) object checksums.

Owner

The owner of a specific resource.

ProjectTeam

Represents the Viewers, Editors, or Owners of a given project.

QueryWriteStatusRequest

Request object for QueryWriteStatus.

QueryWriteStatusResponse

Response object for QueryWriteStatus.

ReadObjectRequest

Request message for ReadObject.

ReadObjectResponse

Response message for ReadObject.

RestoreObjectRequest

Message for restoring an object. bucket, object, and generation must be set.

RewriteObjectRequest

Request message for RewriteObject. If the source object is encrypted using a Customer-Supplied Encryption Key the key information must be provided in the copy_source_encryption_algorithm, copy_source_encryption_key_bytes, and copy_source_encryption_key_sha256_bytes fields. If the destination object should be encrypted the keying information should be provided in the encryption_algorithm, encryption_key_bytes, and encryption_key_sha256_bytes fields of the common_object_request_params.customer_encryption field.

RewriteResponse

A rewrite response.

ServiceAccount

A service account, owned by Cloud Storage, which may be used when taking action on behalf of a given project, for example to publish Pub/Sub notifications or to retrieve security keys.

ServiceConstants

Shared constants.

StartResumableWriteRequest

Request message StartResumableWrite.

StartResumableWriteResponse

Response object for StartResumableWrite.

UpdateBucketRequest

Request for UpdateBucket method.

UpdateHmacKeyRequest

Request object to update an HMAC key state. HmacKeyMetadata.state is required and the only writable field in UpdateHmacKey operation. Specifying fields other than state will result in an error.

UpdateObjectRequest

Request message for UpdateObject.

WriteObjectRequest

Request message for WriteObject.

WriteObjectResponse

Response message for WriteObject.

WriteObjectSpec

Describes an attempt to insert an object, possibly over multiple requests.