Module google_api_proto::google::storagetransfer::v1

Modules

agent_pool

Nested message and enum types in AgentPool.

aws_s3_compatible_data

Nested message and enum types in AwsS3CompatibleData.

aws_s3_data

Nested message and enum types in AwsS3Data.

logging_config

Nested message and enum types in LoggingConfig.

metadata_options

Nested message and enum types in MetadataOptions.

notification_config

Nested message and enum types in NotificationConfig.

s3_compatible_metadata

Nested message and enum types in S3CompatibleMetadata.

storage_transfer_service_client

Generated client implementations.

transfer_job

Nested message and enum types in TransferJob.

transfer_operation

Nested message and enum types in TransferOperation.

transfer_options

Nested message and enum types in TransferOptions.

transfer_spec

Nested message and enum types in TransferSpec.

Structs

AgentPool

Represents an agent pool.

AwsAccessKey

AWS access key (see AWS Security Credentials).

AwsS3CompatibleData

An AwsS3CompatibleData resource.

AwsS3Data

An AwsS3Data resource can be a data source, but not a data sink. In an AwsS3Data resource, an object’s name is the S3 object’s key name.

AzureBlobStorageData

An AzureBlobStorageData resource can be a data source, but not a data sink. An AzureBlobStorageData resource represents one Azure container. The storage account determines the Azure endpoint. In an AzureBlobStorageData resource, a blobs’s name is the Azure Blob Storage blob’s key name.

AzureCredentials

Azure credentials

CreateAgentPoolRequest

Specifies the request passed to CreateAgentPool.

CreateTransferJobRequest

Request passed to CreateTransferJob.

DeleteAgentPoolRequest

Specifies the request passed to DeleteAgentPool.

DeleteTransferJobRequest

Request passed to DeleteTransferJob.

ErrorLogEntry

An entry describing an error that has occurred.

ErrorSummary

A summary of errors by error code, plus a count and sample error log entries.

EventStream

Specifies the Event-driven transfer options. Event-driven transfers listen to an event stream to transfer updated files.

GcsData

In a GcsData resource, an object’s name is the Cloud Storage object’s name and its “last modification time” refers to the object’s updated property of Cloud Storage objects, which changes when the content or the metadata of the object is updated.

GetAgentPoolRequest

Specifies the request passed to GetAgentPool.

GetGoogleServiceAccountRequest

Request passed to GetGoogleServiceAccount.

GetTransferJobRequest

Request passed to GetTransferJob.

GoogleServiceAccount

Google service account

HdfsData

An HdfsData resource specifies a path within an HDFS entity (e.g. a cluster). All cluster-specific settings, such as namenodes and ports, are configured on the transfer agents servicing requests, so HdfsData only contains the root path to the data in our transfer.

HttpData

An HttpData resource specifies a list of objects on the web to be transferred over HTTP. The information of the objects to be transferred is contained in a file referenced by a URL. The first line in the file must be "TsvHttpData-1.0", which specifies the format of the file. Subsequent lines specify the information of the list of objects, one object per list entry. Each entry has the following tab-delimited fields:

ListAgentPoolsRequest

The request passed to ListAgentPools.

ListAgentPoolsResponse

Response from ListAgentPools.

ListTransferJobsRequest

projectId, jobNames, and jobStatuses are query parameters that can be specified when listing transfer jobs.

ListTransferJobsResponse

Response from ListTransferJobs.

LoggingConfig

Specifies the logging behavior for transfer operations.

MetadataOptions

Specifies the metadata options for running a transfer.

NotificationConfig

Specification to configure notifications published to Pub/Sub. Notifications are published to the customer-provided topic using the following PubsubMessage.attributes:

ObjectConditions

Conditions that determine which objects are transferred. Applies only to Cloud Data Sources such as S3, Azure, and Cloud Storage.

PauseTransferOperationRequest

Request passed to PauseTransferOperation.

PosixFilesystem

A POSIX filesystem resource.

ResumeTransferOperationRequest

Request passed to ResumeTransferOperation.

RunTransferJobRequest

Request passed to RunTransferJob.

S3CompatibleMetadata

S3CompatibleMetadata contains the metadata fields that apply to the basic types of S3-compatible data providers.

Schedule

Transfers can be scheduled to recur or to run just once.

TransferCounters

A collection of counters that report the progress of a transfer operation.

TransferJob

This resource represents the configuration of a transfer job that runs periodically.

TransferManifest

Specifies where the manifest is located.

TransferOperation

A description of the execution of a transfer.

TransferOptions

TransferOptions define the actions to be performed on objects in a transfer.

TransferSpec

Configuration for running a transfer.

UpdateAgentPoolRequest

Specifies the request passed to UpdateAgentPool.

UpdateTransferJobRequest

Request passed to UpdateTransferJob.