Struct google_api_proto::google::streetview::publish::v1::PhotoSequence

pub struct PhotoSequence {

    pub id: String,
    pub photos: Vec<Photo>,
    pub upload_reference: Option<UploadRef>,
    pub capture_time_override: Option<Timestamp>,
    pub upload_time: Option<Timestamp>,
    pub raw_gps_timeline: Vec<Pose>,
    pub gps_source: i32,
    pub imu: Option<Imu>,
    pub processing_state: i32,
    pub failure_reason: i32,
    pub failure_details: Option<ProcessingFailureDetails>,
    pub distance_meters: f64,
    pub sequence_bounds: Option<LatLngBounds>,
    pub view_count: i64,
    pub filename: String,

}

A sequence of 360 photos along with metadata.

Fields
id: String

Output only. Unique identifier for the photo sequence. This also acts as a long running operation ID if uploading is performed asynchronously.

photos: Vec<Photo>

Output only. Photos with increasing timestamps.

upload_reference: Option<UploadRef>

Input only. Required when creating photo sequence. The resource name where the bytes of the photo sequence (in the form of video) are uploaded.

capture_time_override: Option<Timestamp>

Optional. Absolute time when the photo sequence starts to be captured. If the photo sequence is a video, this is the start time of the video. If this field is populated in input, it overrides the capture time in the video or XDM file.

upload_time: Option<Timestamp>

Output only. The time this photo sequence was created in uSV Store service.

raw_gps_timeline: Vec<Pose>

Input only. Raw GPS measurements with increasing timestamps from the device that aren’t time synced with each photo. These raw measurements will be used to infer the pose of each frame. Required in input when InputType is VIDEO and raw GPS measurements are not in Camera Motion Metadata Track (CAMM). User can indicate which takes precedence using gps_source if raw GPS measurements are provided in both raw_gps_timeline and Camera Motion Metadata Track (CAMM).

gps_source: i32

Input only. If both raw_gps_timeline and the Camera Motion Metadata Track (CAMM) contain GPS measurements, indicate which takes precedence.

imu: Option<Imu>

Input only. Three axis IMU data for the collection. If this data is too large to put in the request, then it should be put in the CAMM track for the video. This data always takes precedence over the equivalent CAMM data, if it exists.

processing_state: i32

Output only. The processing state of this sequence.

failure_reason: i32

Output only. If this sequence has processing_state = FAILED, this will contain the reason why it failed. If the processing_state is any other value, this field will be unset.

failure_details: Option<ProcessingFailureDetails>

Output only. If this sequence has failure_reason set, this may contain additional details about the failure.

distance_meters: f64

Output only. The computed distance of the photo sequence in meters.

sequence_bounds: Option<LatLngBounds>

Output only. A rectangular box that encapsulates every image in this photo sequence.

view_count: i64

Output only. The total number of views that all the published images in this PhotoSequence have received.

filename: String

Output only. The filename of the upload. Does not include the directory path. Only available if the sequence was uploaded on a platform that provides the filename.

Implementations

impl PhotoSequence

pub fn gps_source(&self) -> GpsSource

Returns the enum value of gps_source, or the default if the field is set to an invalid enum value.

pub fn set_gps_source(&mut self, value: GpsSource)

Sets gps_source to the provided enum value.

pub fn processing_state(&self) -> ProcessingState

Returns the enum value of processing_state, or the default if the field is set to an invalid enum value.

pub fn set_processing_state(&mut self, value: ProcessingState)

Sets processing_state to the provided enum value.

pub fn failure_reason(&self) -> ProcessingFailureReason

Returns the enum value of failure_reason, or the default if the field is set to an invalid enum value.

pub fn set_failure_reason(&mut self, value: ProcessingFailureReason)

Sets failure_reason to the provided enum value.