Struct google_api_proto::grafeas::v1::grafeas_client::GrafeasClient

pub struct GrafeasClient<T> { /* private fields */ }

Grafeas API.

Retrieves analysis results of Cloud components such as Docker container images.

Analysis results are stored as a series of occurrences. An Occurrence contains information about a specific analysis instance on a resource. An occurrence refers to a Note. A note contains details describing the analysis and is generally stored in a separate project, called a Provider. Multiple occurrences can refer to the same note.

For example, an SSL vulnerability could affect multiple images. In this case, there would be one note for the vulnerability and an occurrence for each image with the vulnerability referring to that note.

Implementations

impl<T> GrafeasClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> GrafeasClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_occurrence( &mut self, request: impl IntoRequest<GetOccurrenceRequest> ) -> Result<Response<Occurrence>, Status>

Gets the specified occurrence.

pub async fn list_occurrences( &mut self, request: impl IntoRequest<ListOccurrencesRequest> ) -> Result<Response<ListOccurrencesResponse>, Status>

Lists occurrences for the specified project.

pub async fn delete_occurrence( &mut self, request: impl IntoRequest<DeleteOccurrenceRequest> ) -> Result<Response<()>, Status>

Deletes the specified occurrence. For example, use this method to delete an occurrence when the occurrence is no longer applicable for the given resource.

pub async fn create_occurrence( &mut self, request: impl IntoRequest<CreateOccurrenceRequest> ) -> Result<Response<Occurrence>, Status>

Creates a new occurrence.

pub async fn batch_create_occurrences( &mut self, request: impl IntoRequest<BatchCreateOccurrencesRequest> ) -> Result<Response<BatchCreateOccurrencesResponse>, Status>

Creates new occurrences in batch.

pub async fn update_occurrence( &mut self, request: impl IntoRequest<UpdateOccurrenceRequest> ) -> Result<Response<Occurrence>, Status>

Updates the specified occurrence.

pub async fn get_occurrence_note( &mut self, request: impl IntoRequest<GetOccurrenceNoteRequest> ) -> Result<Response<Note>, Status>

Gets the note attached to the specified occurrence. Consumer projects can use this method to get a note that belongs to a provider project.

pub async fn get_note( &mut self, request: impl IntoRequest<GetNoteRequest> ) -> Result<Response<Note>, Status>

Gets the specified note.

pub async fn list_notes( &mut self, request: impl IntoRequest<ListNotesRequest> ) -> Result<Response<ListNotesResponse>, Status>

Lists notes for the specified project.

pub async fn delete_note( &mut self, request: impl IntoRequest<DeleteNoteRequest> ) -> Result<Response<()>, Status>

Deletes the specified note.

pub async fn create_note( &mut self, request: impl IntoRequest<CreateNoteRequest> ) -> Result<Response<Note>, Status>

Creates a new note.

pub async fn batch_create_notes( &mut self, request: impl IntoRequest<BatchCreateNotesRequest> ) -> Result<Response<BatchCreateNotesResponse>, Status>

Creates new notes in batch.

pub async fn update_note( &mut self, request: impl IntoRequest<UpdateNoteRequest> ) -> Result<Response<Note>, Status>

Updates the specified note.

pub async fn list_note_occurrences( &mut self, request: impl IntoRequest<ListNoteOccurrencesRequest> ) -> Result<Response<ListNoteOccurrencesResponse>, Status>

Lists occurrences referencing the specified note. Provider projects can use this method to get all occurrences across consumer projects referencing the specified note.