Struct google_api_proto::grafeas::v1::vulnerability_note::Detail

pub struct Detail {

    pub severity_name: String,
    pub description: String,
    pub package_type: String,
    pub affected_cpe_uri: String,
    pub affected_package: String,
    pub affected_version_start: Option<Version>,
    pub affected_version_end: Option<Version>,
    pub fixed_cpe_uri: String,
    pub fixed_package: String,
    pub fixed_version: Option<Version>,
    pub is_obsolete: bool,
    pub source_update_time: Option<Timestamp>,
    pub source: String,
    pub vendor: String,

}

A detail for a distro and package affected by this vulnerability and its associated fix (if one is available).

Fields
severity_name: String

The distro assigned severity of this vulnerability.

description: String

A vendor-specific description of this vulnerability.

package_type: String

The type of package; whether native or non native (e.g., ruby gems, node.js packages, etc.).

affected_cpe_uri: String

Required. The CPE URI this vulnerability affects.

affected_package: String

Required. The package this vulnerability affects.

affected_version_start: Option<Version>

The version number at the start of an interval in which this vulnerability exists. A vulnerability can affect a package between version numbers that are disjoint sets of intervals (example: [1.0.0-1.1.0], [2.4.6-2.4.8] and [4.5.6-4.6.8]) each of which will be represented in its own Detail. If a specific affected version is provided by a vulnerability database, affected_version_start and affected_version_end will be the same in that Detail.

affected_version_end: Option<Version>

The version number at the end of an interval in which this vulnerability exists. A vulnerability can affect a package between version numbers that are disjoint sets of intervals (example: [1.0.0-1.1.0], [2.4.6-2.4.8] and [4.5.6-4.6.8]) each of which will be represented in its own Detail. If a specific affected version is provided by a vulnerability database, affected_version_start and affected_version_end will be the same in that Detail.

fixed_cpe_uri: String

The distro recommended CPE URI to update to that contains a fix for this vulnerability. It is possible for this to be different from the affected_cpe_uri.

fixed_package: String

The distro recommended package to update to that contains a fix for this vulnerability. It is possible for this to be different from the affected_package.

fixed_version: Option<Version>

The distro recommended version to update to that contains a fix for this vulnerability. Setting this to VersionKind.MAXIMUM means no such version is yet available.

is_obsolete: bool

Whether this detail is obsolete. Occurrences are expected not to point to obsolete details.

source_update_time: Option<Timestamp>

The time this information was last changed at the source. This is an upstream timestamp from the underlying information source - e.g. Ubuntu security tracker.

source: String

The source from which the information in this Detail was obtained.

vendor: String

The name of the vendor of the product.