Module google_api_proto::grafeas::v1beta1::attestation::pgp_signed_attestation

Nested message and enum types in PgpSignedAttestation.

Enums

ContentType

Type (for example schema) of the attestation payload that was signed.

KeyId

This field is used by verifiers to select the public key used to validate the signature. Note that the policy of the verifier ultimately determines which public keys verify a signature based on the context of the verification. There is no guarantee validation will succeed if the verifier has no key matching this ID, even if it has a key under a different ID that would verify the signature. Note that this ID should also be present in the signature content above, but that is not expected to be used by the verifier.