Module google_api_proto::grafeas::v1beta1::provenance

Modules

hash

Nested message and enum types in Hash.

Structs

Artifact

Artifact describes a build product.

BuildProvenance

Provenance of a build. Contains all information needed to verify the full details about the build from source to completion.

Command

Command describes a step performed as part of the build pipeline.

FileHashes

Container message for hashes of byte content of files, used in source messages to verify integrity of source input to the build.

Hash

Container message for hash values.

Source

Source describes the location of the source used for the build.