Module google_api_proto::maps::fleetengine::v1

Modules

report_billable_trip_request

Nested message and enum types in ReportBillableTripRequest.

request_header

Nested message and enum types in RequestHeader.

search_vehicles_request

Nested message and enum types in SearchVehiclesRequest.

speed_reading_interval

Nested message and enum types in SpeedReadingInterval.

terminal_point_id

Nested message and enum types in TerminalPointId.

trip_service_client

Generated client implementations.

vehicle

Nested message and enum types in Vehicle.

vehicle_attribute

Nested message and enum types in VehicleAttribute.

vehicle_match

Nested message and enum types in VehicleMatch.

vehicle_service_client

Generated client implementations.

visual_traffic_report_polyline_rendering

Nested message and enum types in VisualTrafficReportPolylineRendering.

Structs

BatteryInfo

Information about the device’s battery.

ConsumableTrafficPolyline

Traffic density along a Vehicle’s path.

CreateTripRequest

CreateTrip request message.

CreateVehicleRequest

CreateVehicle request message.

DeviceSettings

Information about various settings on the mobile device.

GetTripRequest

GetTrip request message.

GetVehicleRequest

GetVehicle request message.

LicensePlate

The license plate information of the Vehicle. To avoid storing personally-identifiable information, only the minimum information about the license plate is stored as part of the entity.

ListVehiclesRequest

ListVehicles request message.

ListVehiclesResponse

ListVehicles response message.

ReportBillableTripRequest

ReportBillableTrip request message.

RequestHeader

A RequestHeader contains fields common to all Fleet Engine RPC requests.

SearchTripsRequest

SearchTrips request message.

SearchTripsResponse

SearchTrips response message.

SearchVehiclesRequest

SearchVehicles request message.

SearchVehiclesResponse

SearchVehicles response message.

SpeedReadingInterval

Traffic density indicator on a contiguous segment of a path. Given a path with points P_0, P_1, … , P_N (zero-based index), the SpeedReadingInterval defines an interval and describes its traffic using the following categories.

StopLocation

The actual location where a stop (pickup/dropoff) happened.

TerminalLocation

Describes the location of a waypoint.

TerminalPointId

Deprecated: TerminalPoints are no longer supported in Fleet Engine. Use TerminalLocation.point instead.

TrafficPolylineData

Traffic conditions along the expected vehicle route.

Trip

Trip metadata.

TripWaypoint

Describes a stopping point on a vehicle’s route or an ending point on a vehicle’s trip.

UpdateTripRequest

UpdateTrip request message.

UpdateVehicleAttributesRequest

UpdateVehicleAttributes request message.

UpdateVehicleAttributesResponse

UpdateVehicleAttributes response message.

UpdateVehicleRequest

`UpdateVehicle request message.

Vehicle

Vehicle metadata.

VehicleAttribute

Describes a vehicle attribute as a key-value pair. The “key:value” string length cannot exceed 256 characters.

VehicleAttributeList

A list-of-lists datatype for vehicle attributes.

VehicleLocation

The location, speed, and heading of a vehicle at a point in time.

VehicleMatch

Contains the vehicle and related estimates for a vehicle that match the points of active trips for the vehicle SearchVehiclesRequest.

VisualTrafficReportPolylineRendering

Describes how clients should color one portion of the polyline along the route.

Waypoint

Describes intermediate points along a route for a VehicleMatch in a SearchVehiclesResponse. This concept is represented as a TripWaypoint in all other endpoints.

Enums

BatteryStatus

Status of the battery, whether full or charging etc.

BillingPlatformIdentifier

A set of values that indicate upon which platform the request was issued.

LocationPowerSaveMode

How location features are configured to behave on the mobile device when the devices “battery saver” feature is on. (https://developer.android.com/reference/android/os/PowerManager#getLocationPowerSaveMode())

LocationSensor

The sensor or methodology used to determine the location.

NavigationStatus

The vehicle’s navigation status.

PolylineFormatType

The type of polyline format.

PowerSource

Type of the charger being used to charge the battery.

TripStatus

The status of a trip indicating its progression.

TripType

The type of a trip.

TripView

Selector for different sets of Trip fields in a GetTrip response. See AIP-157 for context. Additional views are likely to be added.

VehicleState

The state of a Vehicle.

WaypointType

The type of waypoint.