Struct google_api_proto::maps::fleetengine::v1::Trip

pub struct Trip {

    pub name: String,
    pub vehicle_id: String,
    pub trip_status: i32,
    pub trip_type: i32,
    pub pickup_point: Option<TerminalLocation>,
    pub actual_pickup_point: Option<StopLocation>,
    pub actual_pickup_arrival_point: Option<StopLocation>,
    pub pickup_time: Option<Timestamp>,
    pub intermediate_destinations: Vec<TerminalLocation>,
    pub intermediate_destinations_version: Option<Timestamp>,
    pub intermediate_destination_index: i32,
    pub actual_intermediate_destination_arrival_points: Vec<StopLocation>,
    pub actual_intermediate_destinations: Vec<StopLocation>,
    pub dropoff_point: Option<TerminalLocation>,
    pub actual_dropoff_point: Option<StopLocation>,
    pub dropoff_time: Option<Timestamp>,
    pub remaining_waypoints: Vec<TripWaypoint>,
    pub vehicle_waypoints: Vec<TripWaypoint>,
    pub route: Vec<LatLng>,
    pub current_route_segment: String,
    pub current_route_segment_version: Option<Timestamp>,
    pub current_route_segment_traffic: Option<ConsumableTrafficPolyline>,
    pub current_route_segment_traffic_version: Option<Timestamp>,
    pub current_route_segment_end_point: Option<TripWaypoint>,
    pub remaining_distance_meters: Option<i32>,
    pub eta_to_first_waypoint: Option<Timestamp>,
    pub remaining_time_to_first_waypoint: Option<Duration>,
    pub remaining_waypoints_version: Option<Timestamp>,
    pub remaining_waypoints_route_version: Option<Timestamp>,
    pub number_of_passengers: i32,
    pub last_location: Option<VehicleLocation>,
    pub last_location_snappable: bool,
    pub view: i32,

}

Trip metadata.

Fields
name: String

Output only. In the format “providers/{provider}/trips/{trip}”

vehicle_id: String

ID of the vehicle making this trip.

trip_status: i32

Current status of the trip.

trip_type: i32

The type of the trip.

pickup_point: Option<TerminalLocation>

Location where customer indicates they will be picked up.

actual_pickup_point: Option<StopLocation>

Input only. The actual location when and where customer was picked up. This field is for provider to provide feedback on actual pickup information.

actual_pickup_arrival_point: Option<StopLocation>

Input only. The actual time and location of the driver arrival at the pickup point. This field is for provider to provide feedback on actual arrival information at the pickup point.

pickup_time: Option<Timestamp>

Output only. Either the estimated future time when the rider(s) will be picked up, or the actual time when they were picked up.

intermediate_destinations: Vec<TerminalLocation>

Intermediate stops in order that the trip requests (in addition to pickup and dropoff). Initially this will not be supported for shared trips.

intermediate_destinations_version: Option<Timestamp>

Indicates the last time the intermediate_destinations was modified. Your server should cache this value and pass it in UpdateTripRequest when update intermediate_destination_index to ensure the intermediate_destinations is not changed.

intermediate_destination_index: i32

When TripStatus is ENROUTE_TO_INTERMEDIATE_DESTINATION, a number between [0..N-1] indicating which intermediate destination the vehicle will cross next. When TripStatus is ARRIVED_AT_INTERMEDIATE_DESTINATION, a number between [0..N-1] indicating which intermediate destination the vehicle is at. The provider sets this value. If there are no intermediate_destinations, this field is ignored.

actual_intermediate_destination_arrival_points: Vec<StopLocation>

Input only. The actual time and location of the driver’s arrival at an intermediate destination. This field is for provider to provide feedback on actual arrival information at intermediate destinations.

actual_intermediate_destinations: Vec<StopLocation>

Input only. The actual time and location when and where the customer was picked up from an intermediate destination. This field is for provider to provide feedback on actual pickup information at intermediate destinations.

dropoff_point: Option<TerminalLocation>

Location where customer indicates they will be dropped off.

actual_dropoff_point: Option<StopLocation>

Input only. The actual time and location when and where customer was dropped off. This field is for provider to provide feedback on actual dropoff information.

dropoff_time: Option<Timestamp>

Output only. Either the estimated future time when the rider(s) will be dropped off at the final destination, or the actual time when they were dropped off.

remaining_waypoints: Vec<TripWaypoint>

Output only. The full path from the current location to the dropoff point, inclusive. This path could include waypoints from other trips.

vehicle_waypoints: Vec<TripWaypoint>

This field supports manual ordering of the waypoints for the trip. It contains all of the remaining waypoints for the assigned vehicle, as well as the pickup and drop-off waypoints for this trip. If the trip hasn’t been assigned to a vehicle, then Fleet Engine ignores this field. For privacy reasons, this field is only populated by the server on UpdateTrip and CreateTrip calls, NOT on GetTrip calls.

route: Vec<LatLng>

Output only. Anticipated route for this trip to the first entry in remaining_waypoints. Note that the first waypoint may belong to a different trip.

current_route_segment: String

Output only. An encoded path to the next waypoint.

Note: This field is intended only for use by the Driver SDK and Consumer SDK. Decoding is not yet supported.

current_route_segment_version: Option<Timestamp>

Output only. Indicates the last time the route was modified.

Note: This field is intended only for use by the Driver SDK and Consumer SDK.

current_route_segment_traffic: Option<ConsumableTrafficPolyline>

Output only. Indicates the traffic conditions along the current_route_segment when they’re available.

Note: This field is intended only for use by the Driver SDK and Consumer SDK.

current_route_segment_traffic_version: Option<Timestamp>

Output only. Indicates the last time the current_route_segment_traffic was modified.

Note: This field is intended only for use by the Driver SDK and Consumer SDK.

current_route_segment_end_point: Option<TripWaypoint>

Output only. The waypoint where current_route_segment ends.

remaining_distance_meters: Option<i32>

Output only. The remaining driving distance in the current_route_segment field. The value is unspecified if the trip is not assigned to a vehicle, or the trip is completed or cancelled.

eta_to_first_waypoint: Option<Timestamp>

Output only. The ETA to the next waypoint (the first entry in the remaining_waypoints field). The value is unspecified if the trip is not assigned to a vehicle, or the trip is inactive (completed or cancelled).

remaining_time_to_first_waypoint: Option<Duration>

Output only. The duration from when the Trip data is returned to the time in Trip.eta_to_first_waypoint. The value is unspecified if the trip is not assigned to a vehicle, or the trip is inactive (completed or cancelled).

remaining_waypoints_version: Option<Timestamp>

Output only. Indicates the last time that remaining_waypoints was changed (a waypoint was added, removed, or changed).

remaining_waypoints_route_version: Option<Timestamp>

Output only. Indicates the last time the remaining_waypoints.path_to_waypoint and remaining_waypoints.traffic_to_waypoint were modified. Your client app should cache this value and pass it in GetTripRequest to ensure the paths and traffic for remaining_waypoints are only returned if updated.

number_of_passengers: i32

Immutable. Indicates the number of passengers on this trip and does not include the driver. A vehicle must have available capacity to be returned in a SearchVehicles response.

last_location: Option<VehicleLocation>

Output only. Indicates the last reported location of the vehicle along the route.

last_location_snappable: bool

Output only. Indicates whether the vehicle’s last_location can be snapped to the current_route_segment. False if last_location or current_route_segment doesn’t exist. It is computed by Fleet Engine. Any update from clients will be ignored.

view: i32

The subset of Trip fields that are populated and how they should be interpreted.

Implementations

impl Trip

pub fn trip_status(&self) -> TripStatus

Returns the enum value of trip_status, or the default if the field is set to an invalid enum value.

pub fn set_trip_status(&mut self, value: TripStatus)

Sets trip_status to the provided enum value.

pub fn trip_type(&self) -> TripType

Returns the enum value of trip_type, or the default if the field is set to an invalid enum value.

pub fn set_trip_type(&mut self, value: TripType)

Sets trip_type to the provided enum value.

pub fn view(&self) -> TripView

Returns the enum value of view, or the default if the field is set to an invalid enum value.

pub fn set_view(&mut self, value: TripView)

Sets view to the provided enum value.