Module google_api_proto::google::ads::admob::v1

Modules

ad_mob_api_client

Generated client implementations.

generate_mediation_report_response

Nested message and enum types in GenerateMediationReportResponse.

generate_network_report_response

Nested message and enum types in GenerateNetworkReportResponse.

mediation_report_spec

Nested message and enum types in MediationReportSpec.

network_report_spec

Nested message and enum types in NetworkReportSpec.

report_row

Nested message and enum types in ReportRow.

report_warning

Nested message and enum types in ReportWarning.

Structs

DateRange

Specification of a single date range. Both dates are inclusive.

GenerateMediationReportRequest

Request to generate an AdMob Mediation report.

GenerateMediationReportResponse

The streaming response for the AdMob Mediation report where the first response contains the report header, then a stream of row responses, and finally a footer as the last response message.

GenerateNetworkReportRequest

Request to generate an AdMob Network report.

GenerateNetworkReportResponse

The streaming response for the AdMob Network report where the first response contains the report header, then a stream of row responses, and finally a footer as the last response message.

GetPublisherAccountRequest

Request to retrieve the specified publisher account.

ListPublisherAccountsRequest

Request to retrieve the AdMob publisher account accessible with the client credential. Currently all credentials have access to at most 1 account.

ListPublisherAccountsResponse

Response for the publisher account list request.

LocalizationSettings

Localization settings for reports, such as currency and language. It affects how metrics are calculated.

MediationReportSpec

The specification for generating an AdMob Mediation report. For example, the specification to get observed ECPM sliced by ad source and app for the ‘US’ and ‘CN’ countries can look like the following example:

NetworkReportSpec

The specification for generating an AdMob Network report. For example, the specification to get clicks and estimated earnings for only the ‘US’ and ‘CN’ countries can look like the following example:

PublisherAccount

A publisher account contains information relevant to the use of this API, such as the time zone used for the reports.

ReportFooter

Groups data available after report generation, for example, warnings and row counts. Always sent as the last message in the stream response.

ReportHeader

Groups data helps to treat the generated report. Always sent as a first message in the stream response.

ReportRow

A row of the returning report.

ReportWarning

Warnings associated with generation of the report.

StringList

List of string values.

Enums

SortOrder

The sorting order.