Module google_api_proto::google::ai::generativelanguage::v1beta

Modules

attribution_source_id

Nested message and enum types in AttributionSourceId.

candidate

Nested message and enum types in Candidate.

chunk

Nested message and enum types in Chunk.

chunk_data

Nested message and enum types in ChunkData.

condition

Nested message and enum types in Condition.

content_filter

Nested message and enum types in ContentFilter.

custom_metadata

Nested message and enum types in CustomMetadata.

dataset

Nested message and enum types in Dataset.

discuss_service_client

Generated client implementations.

generate_answer_request

Nested message and enum types in GenerateAnswerRequest.

generate_answer_response

Nested message and enum types in GenerateAnswerResponse.

generate_content_response

Nested message and enum types in GenerateContentResponse.

generative_service_client

Generated client implementations.

model_service_client

Generated client implementations.

part

Nested message and enum types in Part.

permission

Nested message and enum types in Permission.

permission_service_client

Generated client implementations.

retriever_service_client

Generated client implementations.

safety_rating

Nested message and enum types in SafetyRating.

safety_setting

Nested message and enum types in SafetySetting.

text_service_client

Generated client implementations.

tuned_model

Nested message and enum types in TunedModel.

tuning_example

Nested message and enum types in TuningExample.

Structs

AttributionSourceId

Identifier for the source contributing to this attribution.

BatchCreateChunksRequest

Request to batch create Chunks.

BatchCreateChunksResponse

Response from BatchCreateChunks containing a list of created Chunks.

BatchDeleteChunksRequest

Request to batch delete Chunks.

BatchEmbedContentsRequest

Batch request to get embeddings from the model for a list of prompts.

BatchEmbedContentsResponse

The response to a BatchEmbedContentsRequest.

BatchEmbedTextRequest

Batch request to get a text embedding from the model.

BatchEmbedTextResponse

The response to a EmbedTextRequest.

BatchUpdateChunksRequest

Request to batch update Chunks.

BatchUpdateChunksResponse

Response from BatchUpdateChunks containing a list of updated Chunks.

Blob

Raw media bytes.

Candidate

A response candidate generated from the model.

Chunk

A Chunk is a subpart of a Document that is treated as an independent unit for the purposes of vector representation and storage. A Corpus can have a maximum of 1 million Chunks.

ChunkData

Extracted data that represents the Chunk content.

CitationMetadata

A collection of source attributions for a piece of content.

CitationSource

A citation to a source for a portion of a specific response.

Condition

Filter condition applicable to a single key.

Content

The base structured datatype containing multi-part content of a message.

ContentEmbedding

A list of floats representing an embedding.

ContentFilter

Content filtering metadata associated with processing a single request.

Corpus

A Corpus is a collection of Documents. A project can create up to 5 corpora.

CountMessageTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountMessageTokensResponse

A response from CountMessageTokens.

CountTextTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountTextTokensResponse

A response from CountTextTokens.

CountTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountTokensResponse

A response from CountTokens.

CreateChunkRequest

Request to create a Chunk.

CreateCorpusRequest

Request to create a Corpus.

CreateDocumentRequest

Request to create a Document.

CreatePermissionRequest

Request to create a Permission.

CreateTunedModelMetadata

Metadata about the state and progress of creating a tuned model returned from the long-running operation

CreateTunedModelRequest

Request to create a TunedModel.

CustomMetadata

User provided metadata stored as key-value pairs.

Dataset

Dataset for training or validation.

DeleteChunkRequest

Request to delete a Chunk.

DeleteCorpusRequest

Request to delete a Corpus.

DeleteDocumentRequest

Request to delete a Document.

DeletePermissionRequest

Request to delete the Permission.

DeleteTunedModelRequest

Request to delete a TunedModel.

Document

A Document is a collection of Chunks. A Corpus can have a maximum of 10,000 Documents.

EmbedContentRequest

Request containing the Content for the model to embed.

EmbedContentResponse

The response to an EmbedContentRequest.

EmbedTextRequest

Request to get a text embedding from the model.

EmbedTextResponse

The response to a EmbedTextRequest.

Embedding

A list of floats representing the embedding.

Example

An input/output example used to instruct the Model.

FunctionCall

A predicted FunctionCall returned from the model that contains a string representing the FunctionDeclaration.name with the arguments and their values.

FunctionDeclaration

Structured representation of a function declaration as defined by the OpenAPI 3.03 specification. Included in this declaration are the function name and parameters. This FunctionDeclaration is a representation of a block of code that can be used as a Tool by the model and executed by the client.

FunctionResponse

The result output from a FunctionCall that contains a string representing the FunctionDeclaration.name and a structured JSON object containing any output from the function is used as context to the model. This should contain the result of aFunctionCall made based on model prediction.

GenerateAnswerRequest

Request to generate a grounded answer from the model.

GenerateAnswerResponse

Response from the model for a grounded answer.

GenerateContentRequest

Request to generate a completion from the model.

GenerateContentResponse

Response from the model supporting multiple candidates.

GenerateMessageRequest

Request to generate a message response from the model.

GenerateMessageResponse

The response from the model.

GenerateTextRequest

Request to generate a text completion response from the model.

GenerateTextResponse

The response from the model, including candidate completions.

GenerationConfig

Configuration options for model generation and outputs. Not all parameters may be configurable for every model.

GetChunkRequest

Request for getting information about a specific Chunk.

GetCorpusRequest

Request for getting information about a specific Corpus.

GetDocumentRequest

Request for getting information about a specific Document.

GetModelRequest

Request for getting information about a specific Model.

GetPermissionRequest

Request for getting information about a specific Permission.

GetTunedModelRequest

Request for getting information about a specific Model.

GroundingAttribution

Attribution for a source that contributed to an answer.

GroundingPassage

Passage included inline with a grounding configuration.

GroundingPassages

A repeated list of passages.

Hyperparameters

Hyperparameters controlling the tuning process.

ListChunksRequest

Request for listing Chunks.

ListChunksResponse

Response from ListChunks containing a paginated list of Chunks. The Chunks are sorted by ascending chunk.create_time.

ListCorporaRequest

Request for listing Corpora.

ListCorporaResponse

Response from ListCorpora containing a paginated list of Corpora. The results are sorted by ascending corpus.create_time.

ListDocumentsRequest

Request for listing Documents.

ListDocumentsResponse

Response from ListDocuments containing a paginated list of Documents. The Documents are sorted by ascending document.create_time.

ListModelsRequest

Request for listing all Models.

ListModelsResponse

Response from ListModel containing a paginated list of Models.

ListPermissionsRequest

Request for listing permissions.

ListPermissionsResponse

Response from ListPermissions containing a paginated list of permissions.

ListTunedModelsRequest

Request for listing TunedModels.

ListTunedModelsResponse

Response from ListTunedModels containing a paginated list of Models.

Message

The base unit of structured text.

MessagePrompt

All of the structured input text passed to the model as a prompt.

MetadataFilter

User provided filter to limit retrieval based on Chunk or Document level metadata values. Example (genre = drama OR genre = action): key = “document.custom_metadata.genre” conditions = [{string_value = “drama”, operation = EQUAL}, {string_value = “action”, operation = EQUAL}]

Model

Information about a Generative Language Model.

Part

A datatype containing media that is part of a multi-part Content message.

Permission

Permission resource grants user, group or the rest of the world access to the PaLM API resource (e.g. a tuned model, corpus).

QueryCorpusRequest

Request for querying a Corpus.

QueryCorpusResponse

Response from QueryCorpus containing a list of relevant chunks.

QueryDocumentRequest

Request for querying a Document.

QueryDocumentResponse

Response from QueryDocument containing a list of relevant chunks.

RelevantChunk

The information for a chunk relevant to a query.

SafetyFeedback

Safety feedback for an entire request.

SafetyRating

Safety rating for a piece of content.

SafetySetting

Safety setting, affecting the safety-blocking behavior.

Schema

The Schema object allows the definition of input and output data types. These types can be objects, but also primitives and arrays. Represents a select subset of an OpenAPI 3.0 schema object.

SemanticRetrieverConfig

Configuration for retrieving grounding content from a Corpus or Document created using the Semantic Retriever API.

StringList

User provided string values assigned to a single metadata key.

TextCompletion

Output text returned from a model.

TextPrompt

Text given to the model as a prompt.

Tool

Tool details that the model may use to generate response.

TransferOwnershipRequest

Request to transfer the ownership of the tuned model.

TransferOwnershipResponse

Response from TransferOwnership.

TunedModel

A fine-tuned model created using ModelService.CreateTunedModel.

TunedModelSource

Tuned model as a source for training a new model.

TuningExample

A single example for tuning.

TuningExamples

A set of tuning examples. Can be training or validation data.

TuningSnapshot

Record for a single tuning step.

TuningTask

Tuning tasks that create tuned models.

UpdateChunkRequest

Request to update a Chunk.

UpdateCorpusRequest

Request to update a Corpus.

UpdateDocumentRequest

Request to update a Document.

UpdatePermissionRequest

Request to update the Permission.

UpdateTunedModelRequest

Request to update a TunedModel.

Enums

HarmCategory

The category of a rating.

TaskType

Type of task for which the embedding will be used.

Type

Type contains the list of OpenAPI data types as defined by https://spec.openapis.org/oas/v3.0.3#data-types