Module google_api_proto::google::ai::generativelanguage::v1beta2

Modules

content_filter

Nested message and enum types in ContentFilter.

discuss_service_client

Generated client implementations.

model_service_client

Generated client implementations.

safety_rating

Nested message and enum types in SafetyRating.

safety_setting

Nested message and enum types in SafetySetting.

text_service_client

Generated client implementations.

Structs

CitationMetadata

A collection of source attributions for a piece of content.

CitationSource

A citation to a source for a portion of a specific response.

ContentFilter

Content filtering metadata associated with processing a single request.

CountMessageTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountMessageTokensResponse

A response from CountMessageTokens.

EmbedTextRequest

Request to get a text embedding from the model.

EmbedTextResponse

The response to a EmbedTextRequest.

Embedding

A list of floats representing the embedding.

Example

An input/output example used to instruct the Model.

GenerateMessageRequest

Request to generate a message response from the model.

GenerateMessageResponse

The response from the model.

GenerateTextRequest

Request to generate a text completion response from the model.

GenerateTextResponse

The response from the model, including candidate completions.

GetModelRequest

Request for getting information about a specific Model.

ListModelsRequest

Request for listing all Models.

ListModelsResponse

Response from ListModel containing a paginated list of Models.

Message

The base unit of structured text.

MessagePrompt

All of the structured input text passed to the model as a prompt.

Model

Information about a Generative Language Model.

SafetyFeedback

Safety feedback for an entire request.

SafetyRating

Safety rating for a piece of content.

SafetySetting

Safety setting, affecting the safety-blocking behavior.

TextCompletion

Output text returned from a model.

TextPrompt

Text given to the model as a prompt.

Enums

HarmCategory

The category of a rating.