Module google_api_proto::google::ai::generativelanguage::v1beta3

Modules

content_filter

Nested message and enum types in ContentFilter.

dataset

Nested message and enum types in Dataset.

discuss_service_client

Generated client implementations.

model_service_client

Generated client implementations.

permission

Nested message and enum types in Permission.

permission_service_client

Generated client implementations.

safety_rating

Nested message and enum types in SafetyRating.

safety_setting

Nested message and enum types in SafetySetting.

text_service_client

Generated client implementations.

tuned_model

Nested message and enum types in TunedModel.

tuning_example

Nested message and enum types in TuningExample.

Structs

BatchEmbedTextRequest

Batch request to get a text embedding from the model.

BatchEmbedTextResponse

The response to a EmbedTextRequest.

CitationMetadata

A collection of source attributions for a piece of content.

CitationSource

A citation to a source for a portion of a specific response.

ContentFilter

Content filtering metadata associated with processing a single request.

CountMessageTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountMessageTokensResponse

A response from CountMessageTokens.

CountTextTokensRequest

Counts the number of tokens in the prompt sent to a model.

CountTextTokensResponse

A response from CountTextTokens.

CreatePermissionRequest

Request to create a Permission.

CreateTunedModelMetadata

Metadata about the state and progress of creating a tuned model returned from the long-running operation

CreateTunedModelRequest

Request to create a TunedModel.

Dataset

Dataset for training or validation.

DeletePermissionRequest

Request to delete the Permission.

DeleteTunedModelRequest

Request to delete a TunedModel.

EmbedTextRequest

Request to get a text embedding from the model.

EmbedTextResponse

The response to a EmbedTextRequest.

Embedding

A list of floats representing the embedding.

Example

An input/output example used to instruct the Model.

GenerateMessageRequest

Request to generate a message response from the model.

GenerateMessageResponse

The response from the model.

GenerateTextRequest

Request to generate a text completion response from the model.

GenerateTextResponse

The response from the model, including candidate completions.

GetModelRequest

Request for getting information about a specific Model.

GetPermissionRequest

Request for getting information about a specific Permission.

GetTunedModelRequest

Request for getting information about a specific Model.

Hyperparameters

Hyperparameters controlling the tuning process.

ListModelsRequest

Request for listing all Models.

ListModelsResponse

Response from ListModel containing a paginated list of Models.

ListPermissionsRequest

Request for listing permissions.

ListPermissionsResponse

Response from ListPermissions containing a paginated list of permissions.

ListTunedModelsRequest

Request for listing TunedModels.

ListTunedModelsResponse

Response from ListTunedModels containing a paginated list of Models.

Message

The base unit of structured text.

MessagePrompt

All of the structured input text passed to the model as a prompt.

Model

Information about a Generative Language Model.

Permission

Permission resource grants user, group or the rest of the world access to the PaLM API resource (e.g. a tuned model, file).

SafetyFeedback

Safety feedback for an entire request.

SafetyRating

Safety rating for a piece of content.

SafetySetting

Safety setting, affecting the safety-blocking behavior.

TextCompletion

Output text returned from a model.

TextPrompt

Text given to the model as a prompt.

TransferOwnershipRequest

Request to transfer the ownership of the tuned model.

TransferOwnershipResponse

Response from TransferOwnership.

TunedModel

A fine-tuned model created using ModelService.CreateTunedModel.

TunedModelSource

Tuned model as a source for training a new model.

TuningExample

A single example for tuning.

TuningExamples

A set of tuning examples. Can be training or validatation data.

TuningSnapshot

Record for a single tuning step.

TuningTask

Tuning tasks that create tuned models.

UpdatePermissionRequest

Request to update the Permission.

UpdateTunedModelRequest

Request to update a TunedModel.

Enums

HarmCategory

The category of a rating.