Module google_api_proto::google::analytics::admin::v1alpha

Modules

access_binding

Nested message and enum types in AccessBinding.

access_filter

Nested message and enum types in AccessFilter.

access_filter_expression

Nested message and enum types in AccessFilterExpression.

access_numeric_filter

Nested message and enum types in AccessNumericFilter.

access_order_by

Nested message and enum types in AccessOrderBy.

access_string_filter

Nested message and enum types in AccessStringFilter.

analytics_admin_service_client

Generated client implementations.

attribution_settings

Nested message and enum types in AttributionSettings.

audience

Nested message and enum types in Audience.

audience_dimension_or_metric_filter

Nested message and enum types in AudienceDimensionOrMetricFilter.

audience_event_trigger

Nested message and enum types in AudienceEventTrigger.

audience_filter_clause

Nested message and enum types in AudienceFilterClause.

audience_filter_expression

Nested message and enum types in AudienceFilterExpression.

audience_sequence_filter

Nested message and enum types in AudienceSequenceFilter.

change_history_change

Nested message and enum types in ChangeHistoryChange.

channel_group_filter

Nested message and enum types in ChannelGroupFilter.

channel_group_filter_expression

Nested message and enum types in ChannelGroupFilterExpression.

custom_dimension

Nested message and enum types in CustomDimension.

custom_metric

Nested message and enum types in CustomMetric.

data_retention_settings

Nested message and enum types in DataRetentionSettings.

data_stream

Nested message and enum types in DataStream.

expanded_data_set_filter

Nested message and enum types in ExpandedDataSetFilter.

expanded_data_set_filter_expression

Nested message and enum types in ExpandedDataSetFilterExpression.

matching_condition

Nested message and enum types in MatchingCondition.

numeric_value

Nested message and enum types in NumericValue.

Structs

AccessBetweenFilter

To express that the result needs to be between two numbers (inclusive).

AccessBinding

A binding of a user to a set of roles.

AccessDateRange

A contiguous range of days: startDate, startDate + 1, …, endDate.

AccessDimension

Dimensions are attributes of your data. For example, the dimension userEmail indicates the email of the user that accessed reporting data. Dimension values in report responses are strings.

AccessDimensionHeader

Describes a dimension column in the report. Dimensions requested in a report produce column entries within rows and DimensionHeaders. However, dimensions used exclusively within filters or expressions do not produce columns in a report; correspondingly, those dimensions do not produce headers.

AccessDimensionValue

The value of a dimension.

AccessFilter

An expression to filter dimension or metric values.

AccessFilterExpression

Expresses dimension or metric filters. The fields in the same expression need to be either all dimensions or all metrics.

AccessFilterExpressionList

A list of filter expressions.

AccessInListFilter

The result needs to be in a list of string values.

AccessMetric

The quantitative measurements of a report. For example, the metric accessCount is the total number of data access records.

AccessMetricHeader

Describes a metric column in the report. Visible metrics requested in a report produce column entries within rows and MetricHeaders. However, metrics used exclusively within filters or expressions do not produce columns in a report; correspondingly, those metrics do not produce headers.

AccessMetricValue

The value of a metric.

AccessNumericFilter

Filters for numeric or date values.

AccessOrderBy

Order bys define how rows will be sorted in the response. For example, ordering rows by descending access count is one ordering, and ordering rows by the country string is a different ordering.

AccessQuota

Current state of all quotas for this Analytics property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

AccessQuotaStatus

Current state for a particular quota group.

AccessRow

Access report data for each row.

AccessStringFilter

The filter for strings.

Account

A resource message representing a Google Analytics account.

AccountSummary

A virtual resource representing an overview of an account and all its child GA4 properties.

AcknowledgeUserDataCollectionRequest

Request message for AcknowledgeUserDataCollection RPC.

AcknowledgeUserDataCollectionResponse

Response message for AcknowledgeUserDataCollection RPC.

AdSenseLink

A link between a GA4 Property and an AdSense for Content ad client.

ApproveDisplayVideo360AdvertiserLinkProposalRequest

Request message for ApproveDisplayVideo360AdvertiserLinkProposal RPC.

ApproveDisplayVideo360AdvertiserLinkProposalResponse

Response message for ApproveDisplayVideo360AdvertiserLinkProposal RPC.

ArchiveAudienceRequest

Request message for ArchiveAudience RPC.

ArchiveCustomDimensionRequest

Request message for ArchiveCustomDimension RPC.

ArchiveCustomMetricRequest

Request message for ArchiveCustomMetric RPC.

AttributionSettings

The attribution settings used for a given property. This is a singleton resource.

Audience

A resource message representing a GA4 Audience.

AudienceDimensionOrMetricFilter

A specific filter for a single dimension or metric.

AudienceEventFilter

A filter that matches events of a single event name. If an event parameter is specified, only the subset of events that match both the single event name and the parameter filter expressions match this event filter.

AudienceEventTrigger

Specifies an event to log when a user joins the Audience.

AudienceFilterClause

A clause for defining either a simple or sequence filter. A filter can be inclusive (For example, users satisfying the filter clause are included in the Audience) or exclusive (For example, users satisfying the filter clause are excluded from the Audience).

AudienceFilterExpression

A logical expression of Audience dimension, metric, or event filters.

AudienceFilterExpressionList

A list of Audience filter expressions.

AudienceSequenceFilter

Defines filters that must occur in a specific order for the user to be a member of the Audience.

AudienceSimpleFilter

Defines a simple filter that a user must satisfy to be a member of the Audience.

AuditUserLink

Read-only resource used to summarize a principal’s effective roles.

AuditUserLinksRequest

Request message for AuditUserLinks RPC.

AuditUserLinksResponse

Response message for AuditUserLinks RPC.

BatchCreateAccessBindingsRequest

Request message for BatchCreateAccessBindings RPC.

BatchCreateAccessBindingsResponse

Response message for BatchCreateAccessBindings RPC.

BatchCreateUserLinksRequest

Request message for BatchCreateUserLinks RPC.

BatchCreateUserLinksResponse

Response message for BatchCreateUserLinks RPC.

BatchDeleteAccessBindingsRequest

Request message for BatchDeleteAccessBindings RPC.

BatchDeleteUserLinksRequest

Request message for BatchDeleteUserLinks RPC.

BatchGetAccessBindingsRequest

Request message for BatchGetAccessBindings RPC.

BatchGetAccessBindingsResponse

Response message for BatchGetAccessBindings RPC.

BatchGetUserLinksRequest

Request message for BatchGetUserLinks RPC.

BatchGetUserLinksResponse

Response message for BatchGetUserLinks RPC.

BatchUpdateAccessBindingsRequest

Request message for BatchUpdateAccessBindings RPC.

BatchUpdateAccessBindingsResponse

Response message for BatchUpdateAccessBindings RPC.

BatchUpdateUserLinksRequest

Request message for BatchUpdateUserLinks RPC.

BatchUpdateUserLinksResponse

Response message for BatchUpdateUserLinks RPC.

BigQueryLink

A link between a GA4 Property and BigQuery project.

CancelDisplayVideo360AdvertiserLinkProposalRequest

Request message for CancelDisplayVideo360AdvertiserLinkProposal RPC.

ChangeHistoryChange

A description of a change to a single Google Analytics resource.

ChangeHistoryEvent

A set of changes within a Google Analytics account or its child properties that resulted from the same cause. Common causes would be updates made in the Google Analytics UI, changes from customer support, or automatic Google Analytics system changes.

ChannelGroup

A resource message representing a Channel Group.

ChannelGroupFilter

A specific filter for a single dimension.

ChannelGroupFilterExpression

A logical expression of Channel Group dimension filters.

ChannelGroupFilterExpressionList

A list of Channel Group filter expressions.

ConnectedSiteTag

Configuration for a specific Connected Site Tag.

ConversionEvent

A conversion event in a Google Analytics property.

CreateAccessBindingRequest

Request message for CreateAccessBinding RPC.

CreateAdSenseLinkRequest

Request message to be passed to CreateAdSenseLink method.

CreateAudienceRequest

Request message for CreateAudience RPC.

CreateChannelGroupRequest

Request message for CreateChannelGroup RPC.

CreateConnectedSiteTagRequest

Request message for CreateConnectedSiteTag RPC.

CreateConnectedSiteTagResponse

Response message for CreateConnectedSiteTag RPC.

CreateConversionEventRequest

Request message for CreateConversionEvent RPC

CreateCustomDimensionRequest

Request message for CreateCustomDimension RPC.

CreateCustomMetricRequest

Request message for CreateCustomMetric RPC.

CreateDataStreamRequest

Request message for CreateDataStream RPC.

CreateDisplayVideo360AdvertiserLinkProposalRequest

Request message for CreateDisplayVideo360AdvertiserLinkProposal RPC.

CreateDisplayVideo360AdvertiserLinkRequest

Request message for CreateDisplayVideo360AdvertiserLink RPC.

CreateEventCreateRuleRequest

Request message for CreateEventCreateRule RPC.

CreateExpandedDataSetRequest

Request message for CreateExpandedDataSet RPC.

CreateFirebaseLinkRequest

Request message for CreateFirebaseLink RPC

CreateGoogleAdsLinkRequest

Request message for CreateGoogleAdsLink RPC

CreateMeasurementProtocolSecretRequest

Request message for CreateMeasurementProtocolSecret RPC

CreatePropertyRequest

Request message for CreateProperty RPC.

CreateSearchAds360LinkRequest

Request message for CreateSearchAds360Link RPC.

CreateUserLinkRequest

Request message for CreateUserLink RPC.

CustomDimension

A definition for a CustomDimension.

CustomMetric

A definition for a custom metric.

DataRetentionSettings

Settings values for data retention. This is a singleton resource.

DataSharingSettings

A resource message representing data sharing settings of a Google Analytics account.

DataStream

A resource message representing a data stream.

DeleteAccessBindingRequest

Request message for DeleteAccessBinding RPC.

DeleteAccountRequest

Request message for DeleteAccount RPC.

DeleteAdSenseLinkRequest

Request message to be passed to DeleteAdSenseLink method.

DeleteChannelGroupRequest

Request message for DeleteChannelGroup RPC.

DeleteConnectedSiteTagRequest

Request message for DeleteConnectedSiteTag RPC.

DeleteConversionEventRequest

Request message for DeleteConversionEvent RPC

DeleteDataStreamRequest

Request message for DeleteDataStream RPC.

DeleteDisplayVideo360AdvertiserLinkProposalRequest

Request message for DeleteDisplayVideo360AdvertiserLinkProposal RPC.

DeleteDisplayVideo360AdvertiserLinkRequest

Request message for DeleteDisplayVideo360AdvertiserLink RPC.

DeleteEventCreateRuleRequest

Request message for DeleteEventCreateRule RPC.

DeleteExpandedDataSetRequest

Request message for DeleteExpandedDataSet RPC.

DeleteFirebaseLinkRequest

Request message for DeleteFirebaseLink RPC

DeleteGoogleAdsLinkRequest

Request message for DeleteGoogleAdsLink RPC.

DeleteMeasurementProtocolSecretRequest

Request message for DeleteMeasurementProtocolSecret RPC

DeletePropertyRequest

Request message for DeleteProperty RPC.

DeleteSearchAds360LinkRequest

Request message for DeleteSearchAds360Link RPC.

DeleteUserLinkRequest

Request message for DeleteUserLink RPC.

DisplayVideo360AdvertiserLink

A link between a GA4 property and a Display & Video 360 advertiser.

DisplayVideo360AdvertiserLinkProposal

A proposal for a link between a GA4 property and a Display & Video 360 advertiser.

EnhancedMeasurementSettings

Singleton resource under a web DataStream, configuring measurement of additional site interactions and content.

EventCreateRule

An Event Create Rule defines conditions that will trigger the creation of an entirely new event based upon matched criteria of a source event. Additional mutations of the parameters from the source event can be defined.

ExpandedDataSet

A resource message representing a GA4 ExpandedDataSet.

ExpandedDataSetFilter

A specific filter for a single dimension

ExpandedDataSetFilterExpression

A logical expression of EnhancedDataSet dimension filters.

ExpandedDataSetFilterExpressionList

A list of ExpandedDataSet filter expressions.

FetchAutomatedGa4ConfigurationOptOutRequest

Request for fetching the opt out status for the automated GA4 setup process.

FetchAutomatedGa4ConfigurationOptOutResponse

Response message for fetching the opt out status for the automated GA4 setup process.

FetchConnectedGa4PropertyRequest

Request for looking up GA4 property connected to a UA property.

FetchConnectedGa4PropertyResponse

Response for looking up GA4 property connected to a UA property.

FirebaseLink

A link between a GA4 property and a Firebase project.

GetAccessBindingRequest

Request message for GetAccessBinding RPC.

GetAccountRequest

Request message for GetAccount RPC.

GetAdSenseLinkRequest

Request message to be passed to GetAdSenseLink method.

GetAttributionSettingsRequest

Request message for GetAttributionSettings RPC.

GetAudienceRequest

Request message for GetAudience RPC.

GetBigQueryLinkRequest

Request message for GetBigQueryLink RPC.

GetChannelGroupRequest

Request message for GetChannelGroup RPC.

GetConversionEventRequest

Request message for GetConversionEvent RPC

GetCustomDimensionRequest

Request message for GetCustomDimension RPC.

GetCustomMetricRequest

Request message for GetCustomMetric RPC.

GetDataRetentionSettingsRequest

Request message for GetDataRetentionSettings RPC.

GetDataSharingSettingsRequest

Request message for GetDataSharingSettings RPC.

GetDataStreamRequest

Request message for GetDataStream RPC.

GetDisplayVideo360AdvertiserLinkProposalRequest

Request message for GetDisplayVideo360AdvertiserLinkProposal RPC.

GetDisplayVideo360AdvertiserLinkRequest

Request message for GetDisplayVideo360AdvertiserLink RPC.

GetEnhancedMeasurementSettingsRequest

Request message for GetEnhancedMeasurementSettings RPC.

GetEventCreateRuleRequest

Request message for GetEventCreateRule RPC.

GetExpandedDataSetRequest

Request message for GetExpandedDataSet RPC.

GetGlobalSiteTagRequest

Request message for GetGlobalSiteTag RPC.

GetGoogleSignalsSettingsRequest

Request message for GetGoogleSignalsSettings RPC

GetMeasurementProtocolSecretRequest

Request message for GetMeasurementProtocolSecret RPC.

GetPropertyRequest

Request message for GetProperty RPC.

GetSearchAds360LinkRequest

Request message for GetSearchAds360Link RPC.

GetUserLinkRequest

Request message for GetUserLink RPC.

GlobalSiteTag

Read-only resource with the tag for sending data from a website to a DataStream. Only present for web DataStream resources.

GoogleAdsLink

A link between a GA4 property and a Google Ads account.

GoogleSignalsSettings

Settings values for Google Signals. This is a singleton resource.

GroupingRule

The rules that govern how traffic is grouped into one channel.

LinkProposalStatusDetails

Status information for a link proposal.

ListAccessBindingsRequest

Request message for ListAccessBindings RPC.

ListAccessBindingsResponse

Response message for ListAccessBindings RPC.

ListAccountSummariesRequest

Request message for ListAccountSummaries RPC.

ListAccountSummariesResponse

Response message for ListAccountSummaries RPC.

ListAccountsRequest

Request message for ListAccounts RPC.

ListAccountsResponse

Request message for ListAccounts RPC.

ListAdSenseLinksRequest

Request message to be passed to ListAdSenseLinks method.

ListAdSenseLinksResponse

Response message for ListAdSenseLinks method.

ListAudiencesRequest

Request message for ListAudiences RPC.

ListAudiencesResponse

Response message for ListAudiences RPC.

ListBigQueryLinksRequest

Request message for ListBigQueryLinks RPC.

ListBigQueryLinksResponse

Response message for ListBigQueryLinks RPC

ListChannelGroupsRequest

Request message for ListChannelGroups RPC.

ListChannelGroupsResponse

Response message for ListChannelGroups RPC.

ListConnectedSiteTagsRequest

Request message for ListConnectedSiteTags RPC.

ListConnectedSiteTagsResponse

Response message for ListConnectedSiteTags RPC.

ListConversionEventsRequest

Request message for ListConversionEvents RPC

ListConversionEventsResponse

Response message for ListConversionEvents RPC.

ListCustomDimensionsRequest

Request message for ListCustomDimensions RPC.

ListCustomDimensionsResponse

Response message for ListCustomDimensions RPC.

ListCustomMetricsRequest

Request message for ListCustomMetrics RPC.

ListCustomMetricsResponse

Response message for ListCustomMetrics RPC.

ListDataStreamsRequest

Request message for ListDataStreams RPC.

ListDataStreamsResponse

Response message for ListDataStreams RPC.

ListDisplayVideo360AdvertiserLinkProposalsRequest

Request message for ListDisplayVideo360AdvertiserLinkProposals RPC.

ListDisplayVideo360AdvertiserLinkProposalsResponse

Response message for ListDisplayVideo360AdvertiserLinkProposals RPC.

ListDisplayVideo360AdvertiserLinksRequest

Request message for ListDisplayVideo360AdvertiserLinks RPC.

ListDisplayVideo360AdvertiserLinksResponse

Response message for ListDisplayVideo360AdvertiserLinks RPC.

ListEventCreateRulesRequest

Request message for ListEventCreateRules RPC.

ListEventCreateRulesResponse

Response message for ListEventCreateRules RPC.

ListExpandedDataSetsRequest

Request message for ListExpandedDataSets RPC.

ListExpandedDataSetsResponse

Response message for ListExpandedDataSets RPC.

ListFirebaseLinksRequest

Request message for ListFirebaseLinks RPC

ListFirebaseLinksResponse

Response message for ListFirebaseLinks RPC

ListGoogleAdsLinksRequest

Request message for ListGoogleAdsLinks RPC.

ListGoogleAdsLinksResponse

Response message for ListGoogleAdsLinks RPC.

ListMeasurementProtocolSecretsRequest

Request message for ListMeasurementProtocolSecret RPC

ListMeasurementProtocolSecretsResponse

Response message for ListMeasurementProtocolSecret RPC

ListPropertiesRequest

Request message for ListProperties RPC.

ListPropertiesResponse

Response message for ListProperties RPC.

ListSearchAds360LinksRequest

Request message for ListSearchAds360Links RPC.

ListSearchAds360LinksResponse

Response message for ListSearchAds360Links RPC.

ListUserLinksRequest

Request message for ListUserLinks RPC.

ListUserLinksResponse

Response message for ListUserLinks RPC.

MatchingCondition

Defines a condition for when an Event Edit or Event Creation rule applies to an event.

MeasurementProtocolSecret

A secret value used for sending hits to Measurement Protocol.

NumericValue

To represent a number.

ParameterMutation

Defines an event parameter to mutate.

Property

A resource message representing a Google Analytics GA4 property.

PropertySummary

A virtual resource representing metadata for a GA4 property.

ProvisionAccountTicketRequest

Request message for ProvisionAccountTicket RPC.

ProvisionAccountTicketResponse

Response message for ProvisionAccountTicket RPC.

RunAccessReportRequest

The request for a Data Access Record Report.

RunAccessReportResponse

The customized Data Access Record Report response.

SearchAds360Link

A link between a GA4 property and a Search Ads 360 entity.

SearchChangeHistoryEventsRequest

Request message for SearchChangeHistoryEvents RPC.

SearchChangeHistoryEventsResponse

Response message for SearchAccounts RPC.

SetAutomatedGa4ConfigurationOptOutRequest

Request for setting the opt out status for the automated GA4 setup process.

SetAutomatedGa4ConfigurationOptOutResponse

Response message for setting the opt out status for the automated GA4 setup process.

UpdateAccessBindingRequest

Request message for UpdateAccessBinding RPC.

UpdateAccountRequest

Request message for UpdateAccount RPC.

UpdateAttributionSettingsRequest

Request message for UpdateAttributionSettings RPC

UpdateAudienceRequest

Request message for UpdateAudience RPC.

UpdateChannelGroupRequest

Request message for UpdateChannelGroup RPC.

UpdateCustomDimensionRequest

Request message for UpdateCustomDimension RPC.

UpdateCustomMetricRequest

Request message for UpdateCustomMetric RPC.

UpdateDataRetentionSettingsRequest

Request message for UpdateDataRetentionSettings RPC.

UpdateDataStreamRequest

Request message for UpdateDataStream RPC.

UpdateDisplayVideo360AdvertiserLinkRequest

Request message for UpdateDisplayVideo360AdvertiserLink RPC.

UpdateEnhancedMeasurementSettingsRequest

Request message for UpdateEnhancedMeasurementSettings RPC.

UpdateEventCreateRuleRequest

Request message for UpdateEventCreateRule RPC.

UpdateExpandedDataSetRequest

Request message for UpdateExpandedDataSet RPC.

UpdateGoogleAdsLinkRequest

Request message for UpdateGoogleAdsLink RPC

UpdateGoogleSignalsSettingsRequest

Request message for UpdateGoogleSignalsSettings RPC

UpdateMeasurementProtocolSecretRequest

Request message for UpdateMeasurementProtocolSecret RPC

UpdatePropertyRequest

Request message for UpdateProperty RPC.

UpdateSearchAds360LinkRequest

Request message for UpdateSearchAds360Link RPC.

UpdateUserLinkRequest

Request message for UpdateUserLink RPC.

UserLink

A resource message representing a user’s permissions on an Account or Property resource.

Enums

ActionType

Types of actions that may change a resource.

ActorType

Different kinds of actors that can make changes to Google Analytics resources.

AudienceFilterScope

Specifies how to evaluate users for joining an Audience.

ChangeHistoryResourceType

Types of resources whose changes may be returned from change history.

GoogleSignalsConsent

Consent field of the Google Signals settings.

GoogleSignalsState

Status of the Google Signals settings.

IndustryCategory

The category selected for this property, used for industry benchmarking.

LinkProposalInitiatingProduct

An indication of which product the user initiated a link proposal from.

LinkProposalState

The state of a link proposal resource.

PropertyType

Types of Property resources.

ServiceLevel

Various levels of service for Google Analytics.