Module google_api_proto::google::analytics::admin::v1alpha

Modules

analytics_admin_service_client

Generated client implementations.

change_history_change

Nested message and enum types in ChangeHistoryChange.

custom_dimension

Nested message and enum types in CustomDimension.

custom_metric

Nested message and enum types in CustomMetric.

data_retention_settings

Nested message and enum types in DataRetentionSettings.

data_stream

Nested message and enum types in DataStream.

Structs

Account

A resource message representing a Google Analytics account.

AccountSummary

A virtual resource representing an overview of an account and all its child GA4 properties.

AcknowledgeUserDataCollectionRequest

Request message for AcknowledgeUserDataCollection RPC.

AcknowledgeUserDataCollectionResponse

Response message for AcknowledgeUserDataCollection RPC.

ApproveDisplayVideo360AdvertiserLinkProposalRequest

Request message for ApproveDisplayVideo360AdvertiserLinkProposal RPC.

ApproveDisplayVideo360AdvertiserLinkProposalResponse

Response message for ApproveDisplayVideo360AdvertiserLinkProposal RPC.

ArchiveCustomDimensionRequest

Request message for ArchiveCustomDimension RPC.

ArchiveCustomMetricRequest

Request message for ArchiveCustomMetric RPC.

AuditUserLink

Read-only resource used to summarize a principal’s effective roles.

AuditUserLinksRequest

Request message for AuditUserLinks RPC.

AuditUserLinksResponse

Response message for AuditUserLinks RPC.

BatchCreateUserLinksRequest

Request message for BatchCreateUserLinks RPC.

BatchCreateUserLinksResponse

Response message for BatchCreateUserLinks RPC.

BatchDeleteUserLinksRequest

Request message for BatchDeleteUserLinks RPC.

BatchGetUserLinksRequest

Request message for BatchGetUserLinks RPC.

BatchGetUserLinksResponse

Response message for BatchGetUserLinks RPC.

BatchUpdateUserLinksRequest

Request message for BatchUpdateUserLinks RPC.

BatchUpdateUserLinksResponse

Response message for BatchUpdateUserLinks RPC.

CancelDisplayVideo360AdvertiserLinkProposalRequest

Request message for CancelDisplayVideo360AdvertiserLinkProposal RPC.

ChangeHistoryChange

A description of a change to a single Google Analytics resource.

ChangeHistoryEvent

A set of changes within a Google Analytics account or its child properties that resulted from the same cause. Common causes would be updates made in the Google Analytics UI, changes from customer support, or automatic Google Analytics system changes.

ConversionEvent

A conversion event in a Google Analytics property.

CreateConversionEventRequest

Request message for CreateConversionEvent RPC

CreateCustomDimensionRequest

Request message for CreateCustomDimension RPC.

CreateCustomMetricRequest

Request message for CreateCustomMetric RPC.

CreateDataStreamRequest

Request message for CreateDataStream RPC.

CreateDisplayVideo360AdvertiserLinkProposalRequest

Request message for CreateDisplayVideo360AdvertiserLinkProposal RPC.

CreateDisplayVideo360AdvertiserLinkRequest

Request message for CreateDisplayVideo360AdvertiserLink RPC.

CreateFirebaseLinkRequest

Request message for CreateFirebaseLink RPC

CreateGoogleAdsLinkRequest

Request message for CreateGoogleAdsLink RPC

CreateMeasurementProtocolSecretRequest

Request message for CreateMeasurementProtocolSecret RPC

CreatePropertyRequest

Request message for CreateProperty RPC.

CreateUserLinkRequest

Request message for CreateUserLink RPC.

CustomDimension

A definition for a CustomDimension.

CustomMetric

A definition for a custom metric.

DataRetentionSettings

Settings values for data retention. This is a singleton resource.

DataSharingSettings

A resource message representing data sharing settings of a Google Analytics account.

DataStream

A resource message representing a data stream.

DeleteAccountRequest

Request message for DeleteAccount RPC.

DeleteConversionEventRequest

Request message for DeleteConversionEvent RPC

DeleteDataStreamRequest

Request message for DeleteDataStream RPC.

DeleteDisplayVideo360AdvertiserLinkProposalRequest

Request message for DeleteDisplayVideo360AdvertiserLinkProposal RPC.

DeleteDisplayVideo360AdvertiserLinkRequest

Request message for DeleteDisplayVideo360AdvertiserLink RPC.

DeleteFirebaseLinkRequest

Request message for DeleteFirebaseLink RPC

DeleteGoogleAdsLinkRequest

Request message for DeleteGoogleAdsLink RPC.

DeleteMeasurementProtocolSecretRequest

Request message for DeleteMeasurementProtocolSecret RPC

DeletePropertyRequest

Request message for DeleteProperty RPC.

DeleteUserLinkRequest

Request message for DeleteUserLink RPC.

DisplayVideo360AdvertiserLink

A link between a GA4 property and a Display & Video 360 advertiser.

DisplayVideo360AdvertiserLinkProposal

A proposal for a link between a GA4 property and a Display & Video 360 advertiser.

FirebaseLink

A link between a GA4 property and a Firebase project.

GetAccountRequest

Request message for GetAccount RPC.

GetConversionEventRequest

Request message for GetConversionEvent RPC

GetCustomDimensionRequest

Request message for GetCustomDimension RPC.

GetCustomMetricRequest

Request message for GetCustomMetric RPC.

GetDataRetentionSettingsRequest

Request message for GetDataRetentionSettings RPC.

GetDataSharingSettingsRequest

Request message for GetDataSharingSettings RPC.

GetDataStreamRequest

Request message for GetDataStream RPC.

GetDisplayVideo360AdvertiserLinkProposalRequest

Request message for GetDisplayVideo360AdvertiserLinkProposal RPC.

GetDisplayVideo360AdvertiserLinkRequest

Request message for GetDisplayVideo360AdvertiserLink RPC.

GetGlobalSiteTagRequest

Request message for GetGlobalSiteTag RPC.

GetGoogleSignalsSettingsRequest

Request message for GetGoogleSignalsSettings RPC

GetMeasurementProtocolSecretRequest

Request message for GetMeasurementProtocolSecret RPC.

GetPropertyRequest

Request message for GetProperty RPC.

GetUserLinkRequest

Request message for GetUserLink RPC.

GlobalSiteTag

Read-only resource with the tag for sending data from a website to a DataStream. Only present for web DataStream resources.

GoogleAdsLink

A link between a GA4 property and a Google Ads account.

GoogleSignalsSettings

Settings values for Google Signals. This is a singleton resource.

LinkProposalStatusDetails

Status information for a link proposal.

ListAccountSummariesRequest

Request message for ListAccountSummaries RPC.

ListAccountSummariesResponse

Response message for ListAccountSummaries RPC.

ListAccountsRequest

Request message for ListAccounts RPC.

ListAccountsResponse

Request message for ListAccounts RPC.

ListConversionEventsRequest

Request message for ListConversionEvents RPC

ListConversionEventsResponse

Response message for ListConversionEvents RPC.

ListCustomDimensionsRequest

Request message for ListCustomDimensions RPC.

ListCustomDimensionsResponse

Response message for ListCustomDimensions RPC.

ListCustomMetricsRequest

Request message for ListCustomMetrics RPC.

ListCustomMetricsResponse

Response message for ListCustomMetrics RPC.

ListDataStreamsRequest

Request message for ListDataStreams RPC.

ListDataStreamsResponse

Response message for ListDataStreams RPC.

ListDisplayVideo360AdvertiserLinkProposalsRequest

Request message for ListDisplayVideo360AdvertiserLinkProposals RPC.

ListDisplayVideo360AdvertiserLinkProposalsResponse

Response message for ListDisplayVideo360AdvertiserLinkProposals RPC.

ListDisplayVideo360AdvertiserLinksRequest

Request message for ListDisplayVideo360AdvertiserLinks RPC.

ListDisplayVideo360AdvertiserLinksResponse

Response message for ListDisplayVideo360AdvertiserLinks RPC.

ListFirebaseLinksRequest

Request message for ListFirebaseLinks RPC

ListFirebaseLinksResponse

Response message for ListFirebaseLinks RPC

ListGoogleAdsLinksRequest

Request message for ListGoogleAdsLinks RPC.

ListGoogleAdsLinksResponse

Response message for ListGoogleAdsLinks RPC.

ListMeasurementProtocolSecretsRequest

Request message for ListMeasurementProtocolSecret RPC

ListMeasurementProtocolSecretsResponse

Response message for ListMeasurementProtocolSecret RPC

ListPropertiesRequest

Request message for ListProperties RPC.

ListPropertiesResponse

Response message for ListProperties RPC.

ListUserLinksRequest

Request message for ListUserLinks RPC.

ListUserLinksResponse

Response message for ListUserLinks RPC.

MeasurementProtocolSecret

A secret value used for sending hits to Measurement Protocol.

Property

A resource message representing a Google Analytics GA4 property.

PropertySummary

A virtual resource representing metadata for a GA4 property.

ProvisionAccountTicketRequest

Request message for ProvisionAccountTicket RPC.

ProvisionAccountTicketResponse

Response message for ProvisionAccountTicket RPC.

SearchChangeHistoryEventsRequest

Request message for SearchChangeHistoryEvents RPC.

SearchChangeHistoryEventsResponse

Response message for SearchAccounts RPC.

UpdateAccountRequest

Request message for UpdateAccount RPC.

UpdateCustomDimensionRequest

Request message for UpdateCustomDimension RPC.

UpdateCustomMetricRequest

Request message for UpdateCustomMetric RPC.

UpdateDataRetentionSettingsRequest

Request message for UpdateDataRetentionSettings RPC.

UpdateDataStreamRequest

Request message for UpdateDataStream RPC.

UpdateDisplayVideo360AdvertiserLinkRequest

Request message for UpdateDisplayVideo360AdvertiserLink RPC.

UpdateGoogleAdsLinkRequest

Request message for UpdateGoogleAdsLink RPC

UpdateGoogleSignalsSettingsRequest

Request message for UpdateGoogleSignalsSettings RPC

UpdateMeasurementProtocolSecretRequest

Request message for UpdateMeasurementProtocolSecret RPC

UpdatePropertyRequest

Request message for UpdateProperty RPC.

UpdateUserLinkRequest

Request message for UpdateUserLink RPC.

UserLink

A resource message representing a user’s permissions on an Account or Property resource.

Enums

ActionType

Types of actions that may change a resource.

ActorType

Different kinds of actors that can make changes to Google Analytics resources.

ChangeHistoryResourceType

Types of resources whose changes may be returned from change history.

GoogleSignalsConsent

Consent field of the Google Signals settings (i.e., whether the user has consented to the Google Signals terms of service.)

GoogleSignalsState

Status of the Google Signals settings (i.e., whether this feature has been enabled for the property).

IndustryCategory

The category selected for this property, used for industry benchmarking.

LinkProposalInitiatingProduct

An indication of which product the user initiated a link proposal from.

LinkProposalState

The state of a link proposal resource.

ServiceLevel

Various levels of service for Google Analytics.