Struct google_api_proto::google::analytics::admin::v1alpha::analytics_admin_service_client::AnalyticsAdminServiceClient

pub struct AnalyticsAdminServiceClient<T> { /* private fields */ }

Service Interface for the Analytics Admin API (GA4).

Implementations

impl<T> AnalyticsAdminServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AnalyticsAdminServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_account( &mut self, request: impl IntoRequest<GetAccountRequest> ) -> Result<Response<Account>, Status>

Lookup for a single Account.

pub async fn list_accounts( &mut self, request: impl IntoRequest<ListAccountsRequest> ) -> Result<Response<ListAccountsResponse>, Status>

Returns all accounts accessible by the caller.

Note that these accounts might not currently have GA4 properties. Soft-deleted (ie: “trashed”) accounts are excluded by default. Returns an empty list if no relevant accounts are found.

pub async fn delete_account( &mut self, request: impl IntoRequest<DeleteAccountRequest> ) -> Result<Response<()>, Status>

Marks target Account as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI.

If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, AccessBindings) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found.

pub async fn update_account( &mut self, request: impl IntoRequest<UpdateAccountRequest> ) -> Result<Response<Account>, Status>

Updates an account.

pub async fn provision_account_ticket( &mut self, request: impl IntoRequest<ProvisionAccountTicketRequest> ) -> Result<Response<ProvisionAccountTicketResponse>, Status>

Requests a ticket for creating an account.

pub async fn list_account_summaries( &mut self, request: impl IntoRequest<ListAccountSummariesRequest> ) -> Result<Response<ListAccountSummariesResponse>, Status>

Returns summaries of all accounts accessible by the caller.

pub async fn get_property( &mut self, request: impl IntoRequest<GetPropertyRequest> ) -> Result<Response<Property>, Status>

Lookup for a single “GA4” Property.

pub async fn list_properties( &mut self, request: impl IntoRequest<ListPropertiesRequest> ) -> Result<Response<ListPropertiesResponse>, Status>

Returns child Properties under the specified parent Account.

Only “GA4” properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: “trashed”) properties are excluded by default. Returns an empty list if no relevant properties are found.

pub async fn create_property( &mut self, request: impl IntoRequest<CreatePropertyRequest> ) -> Result<Response<Property>, Status>

Creates an “GA4” property with the specified location and attributes.

pub async fn delete_property( &mut self, request: impl IntoRequest<DeletePropertyRequest> ) -> Result<Response<Property>, Status>

Marks target Property as soft-deleted (ie: “trashed”) and returns it.

This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI.

If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, AccessBindings) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found, or is not a GA4 Property.

pub async fn update_property( &mut self, request: impl IntoRequest<UpdatePropertyRequest> ) -> Result<Response<Property>, Status>

Updates a property.

pub async fn create_firebase_link( &mut self, request: impl IntoRequest<CreateFirebaseLinkRequest> ) -> Result<Response<FirebaseLink>, Status>

Creates a FirebaseLink.

Properties can have at most one FirebaseLink.

pub async fn delete_firebase_link( &mut self, request: impl IntoRequest<DeleteFirebaseLinkRequest> ) -> Result<Response<()>, Status>

Deletes a FirebaseLink on a property

pub async fn list_firebase_links( &mut self, request: impl IntoRequest<ListFirebaseLinksRequest> ) -> Result<Response<ListFirebaseLinksResponse>, Status>

Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

pub async fn get_global_site_tag( &mut self, request: impl IntoRequest<GetGlobalSiteTagRequest> ) -> Result<Response<GlobalSiteTag>, Status>

Returns the Site Tag for the specified web stream. Site Tags are immutable singletons.

pub async fn create_google_ads_link( &mut self, request: impl IntoRequest<CreateGoogleAdsLinkRequest> ) -> Result<Response<GoogleAdsLink>, Status>

Creates a GoogleAdsLink.

pub async fn update_google_ads_link( &mut self, request: impl IntoRequest<UpdateGoogleAdsLinkRequest> ) -> Result<Response<GoogleAdsLink>, Status>

Updates a GoogleAdsLink on a property

pub async fn delete_google_ads_link( &mut self, request: impl IntoRequest<DeleteGoogleAdsLinkRequest> ) -> Result<Response<()>, Status>

Deletes a GoogleAdsLink on a property

pub async fn list_google_ads_links( &mut self, request: impl IntoRequest<ListGoogleAdsLinksRequest> ) -> Result<Response<ListGoogleAdsLinksResponse>, Status>

Lists GoogleAdsLinks on a property.

pub async fn get_data_sharing_settings( &mut self, request: impl IntoRequest<GetDataSharingSettingsRequest> ) -> Result<Response<DataSharingSettings>, Status>

Get data sharing settings on an account. Data sharing settings are singletons.

pub async fn get_measurement_protocol_secret( &mut self, request: impl IntoRequest<GetMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Lookup for a single “GA4” MeasurementProtocolSecret.

pub async fn list_measurement_protocol_secrets( &mut self, request: impl IntoRequest<ListMeasurementProtocolSecretsRequest> ) -> Result<Response<ListMeasurementProtocolSecretsResponse>, Status>

Returns child MeasurementProtocolSecrets under the specified parent Property.

pub async fn create_measurement_protocol_secret( &mut self, request: impl IntoRequest<CreateMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Creates a measurement protocol secret.

pub async fn delete_measurement_protocol_secret( &mut self, request: impl IntoRequest<DeleteMeasurementProtocolSecretRequest> ) -> Result<Response<()>, Status>

Deletes target MeasurementProtocolSecret.

pub async fn update_measurement_protocol_secret( &mut self, request: impl IntoRequest<UpdateMeasurementProtocolSecretRequest> ) -> Result<Response<MeasurementProtocolSecret>, Status>

Updates a measurement protocol secret.

pub async fn acknowledge_user_data_collection( &mut self, request: impl IntoRequest<AcknowledgeUserDataCollectionRequest> ) -> Result<Response<AcknowledgeUserDataCollectionResponse>, Status>

Acknowledges the terms of user data collection for the specified property.

This acknowledgement must be completed (either in the Google Analytics UI or through this API) before MeasurementProtocolSecret resources may be created.

pub async fn get_sk_ad_network_conversion_value_schema( &mut self, request: impl IntoRequest<GetSkAdNetworkConversionValueSchemaRequest> ) -> Result<Response<SkAdNetworkConversionValueSchema>, Status>

Looks up a single SKAdNetworkConversionValueSchema.

pub async fn create_sk_ad_network_conversion_value_schema( &mut self, request: impl IntoRequest<CreateSkAdNetworkConversionValueSchemaRequest> ) -> Result<Response<SkAdNetworkConversionValueSchema>, Status>

Creates a SKAdNetworkConversionValueSchema.

pub async fn delete_sk_ad_network_conversion_value_schema( &mut self, request: impl IntoRequest<DeleteSkAdNetworkConversionValueSchemaRequest> ) -> Result<Response<()>, Status>

Deletes target SKAdNetworkConversionValueSchema.

pub async fn update_sk_ad_network_conversion_value_schema( &mut self, request: impl IntoRequest<UpdateSkAdNetworkConversionValueSchemaRequest> ) -> Result<Response<SkAdNetworkConversionValueSchema>, Status>

Updates a SKAdNetworkConversionValueSchema.

pub async fn list_sk_ad_network_conversion_value_schemas( &mut self, request: impl IntoRequest<ListSkAdNetworkConversionValueSchemasRequest> ) -> Result<Response<ListSkAdNetworkConversionValueSchemasResponse>, Status>

Lists SKAdNetworkConversionValueSchema on a stream. Properties can have at most one SKAdNetworkConversionValueSchema.

pub async fn search_change_history_events( &mut self, request: impl IntoRequest<SearchChangeHistoryEventsRequest> ) -> Result<Response<SearchChangeHistoryEventsResponse>, Status>

Searches through all changes to an account or its children given the specified set of filters.

pub async fn get_google_signals_settings( &mut self, request: impl IntoRequest<GetGoogleSignalsSettingsRequest> ) -> Result<Response<GoogleSignalsSettings>, Status>

Lookup for Google Signals settings for a property.

pub async fn update_google_signals_settings( &mut self, request: impl IntoRequest<UpdateGoogleSignalsSettingsRequest> ) -> Result<Response<GoogleSignalsSettings>, Status>

Updates Google Signals settings for a property.

pub async fn create_conversion_event( &mut self, request: impl IntoRequest<CreateConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Creates a conversion event with the specified attributes.

pub async fn update_conversion_event( &mut self, request: impl IntoRequest<UpdateConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Updates a conversion event with the specified attributes.

pub async fn get_conversion_event( &mut self, request: impl IntoRequest<GetConversionEventRequest> ) -> Result<Response<ConversionEvent>, Status>

Retrieve a single conversion event.

pub async fn delete_conversion_event( &mut self, request: impl IntoRequest<DeleteConversionEventRequest> ) -> Result<Response<()>, Status>

Deletes a conversion event in a property.

pub async fn list_conversion_events( &mut self, request: impl IntoRequest<ListConversionEventsRequest> ) -> Result<Response<ListConversionEventsResponse>, Status>

Returns a list of conversion events in the specified parent property.

Returns an empty list if no conversion events are found.

pub async fn get_display_video360_advertiser_link( &mut self, request: impl IntoRequest<GetDisplayVideo360AdvertiserLinkRequest> ) -> Result<Response<DisplayVideo360AdvertiserLink>, Status>

Look up a single DisplayVideo360AdvertiserLink

pub async fn list_display_video360_advertiser_links( &mut self, request: impl IntoRequest<ListDisplayVideo360AdvertiserLinksRequest> ) -> Result<Response<ListDisplayVideo360AdvertiserLinksResponse>, Status>

Lists all DisplayVideo360AdvertiserLinks on a property.

pub async fn create_display_video360_advertiser_link( &mut self, request: impl IntoRequest<CreateDisplayVideo360AdvertiserLinkRequest> ) -> Result<Response<DisplayVideo360AdvertiserLink>, Status>

Creates a DisplayVideo360AdvertiserLink. This can only be utilized by users who have proper authorization both on the Google Analytics property and on the Display & Video 360 advertiser. Users who do not have access to the Display & Video 360 advertiser should instead seek to create a DisplayVideo360LinkProposal.

pub async fn delete_display_video360_advertiser_link( &mut self, request: impl IntoRequest<DeleteDisplayVideo360AdvertiserLinkRequest> ) -> Result<Response<()>, Status>

Deletes a DisplayVideo360AdvertiserLink on a property.

pub async fn update_display_video360_advertiser_link( &mut self, request: impl IntoRequest<UpdateDisplayVideo360AdvertiserLinkRequest> ) -> Result<Response<DisplayVideo360AdvertiserLink>, Status>

Updates a DisplayVideo360AdvertiserLink on a property.

pub async fn get_display_video360_advertiser_link_proposal( &mut self, request: impl IntoRequest<GetDisplayVideo360AdvertiserLinkProposalRequest> ) -> Result<Response<DisplayVideo360AdvertiserLinkProposal>, Status>

Lookup for a single DisplayVideo360AdvertiserLinkProposal.

pub async fn list_display_video360_advertiser_link_proposals( &mut self, request: impl IntoRequest<ListDisplayVideo360AdvertiserLinkProposalsRequest> ) -> Result<Response<ListDisplayVideo360AdvertiserLinkProposalsResponse>, Status>

Lists DisplayVideo360AdvertiserLinkProposals on a property.

pub async fn create_display_video360_advertiser_link_proposal( &mut self, request: impl IntoRequest<CreateDisplayVideo360AdvertiserLinkProposalRequest> ) -> Result<Response<DisplayVideo360AdvertiserLinkProposal>, Status>

Creates a DisplayVideo360AdvertiserLinkProposal.

pub async fn delete_display_video360_advertiser_link_proposal( &mut self, request: impl IntoRequest<DeleteDisplayVideo360AdvertiserLinkProposalRequest> ) -> Result<Response<()>, Status>

Deletes a DisplayVideo360AdvertiserLinkProposal on a property. This can only be used on cancelled proposals.

pub async fn approve_display_video360_advertiser_link_proposal( &mut self, request: impl IntoRequest<ApproveDisplayVideo360AdvertiserLinkProposalRequest> ) -> Result<Response<ApproveDisplayVideo360AdvertiserLinkProposalResponse>, Status>

Approves a DisplayVideo360AdvertiserLinkProposal. The DisplayVideo360AdvertiserLinkProposal will be deleted and a new DisplayVideo360AdvertiserLink will be created.

pub async fn cancel_display_video360_advertiser_link_proposal( &mut self, request: impl IntoRequest<CancelDisplayVideo360AdvertiserLinkProposalRequest> ) -> Result<Response<DisplayVideo360AdvertiserLinkProposal>, Status>

Cancels a DisplayVideo360AdvertiserLinkProposal. Cancelling can mean either:

Declining a proposal initiated from Display & Video 360

Withdrawing a proposal initiated from Google Analytics After being cancelled, a proposal will eventually be deleted automatically.

pub async fn create_custom_dimension( &mut self, request: impl IntoRequest<CreateCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Creates a CustomDimension.

pub async fn update_custom_dimension( &mut self, request: impl IntoRequest<UpdateCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Updates a CustomDimension on a property.

pub async fn list_custom_dimensions( &mut self, request: impl IntoRequest<ListCustomDimensionsRequest> ) -> Result<Response<ListCustomDimensionsResponse>, Status>

Lists CustomDimensions on a property.

pub async fn archive_custom_dimension( &mut self, request: impl IntoRequest<ArchiveCustomDimensionRequest> ) -> Result<Response<()>, Status>

Archives a CustomDimension on a property.

pub async fn get_custom_dimension( &mut self, request: impl IntoRequest<GetCustomDimensionRequest> ) -> Result<Response<CustomDimension>, Status>

Lookup for a single CustomDimension.

pub async fn create_custom_metric( &mut self, request: impl IntoRequest<CreateCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Creates a CustomMetric.

pub async fn update_custom_metric( &mut self, request: impl IntoRequest<UpdateCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Updates a CustomMetric on a property.

pub async fn list_custom_metrics( &mut self, request: impl IntoRequest<ListCustomMetricsRequest> ) -> Result<Response<ListCustomMetricsResponse>, Status>

Lists CustomMetrics on a property.

pub async fn archive_custom_metric( &mut self, request: impl IntoRequest<ArchiveCustomMetricRequest> ) -> Result<Response<()>, Status>

Archives a CustomMetric on a property.

pub async fn get_custom_metric( &mut self, request: impl IntoRequest<GetCustomMetricRequest> ) -> Result<Response<CustomMetric>, Status>

Lookup for a single CustomMetric.

pub async fn get_data_retention_settings( &mut self, request: impl IntoRequest<GetDataRetentionSettingsRequest> ) -> Result<Response<DataRetentionSettings>, Status>

Returns the singleton data retention settings for this property.

pub async fn update_data_retention_settings( &mut self, request: impl IntoRequest<UpdateDataRetentionSettingsRequest> ) -> Result<Response<DataRetentionSettings>, Status>

Updates the singleton data retention settings for this property.

pub async fn create_data_stream( &mut self, request: impl IntoRequest<CreateDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Creates a DataStream.

pub async fn delete_data_stream( &mut self, request: impl IntoRequest<DeleteDataStreamRequest> ) -> Result<Response<()>, Status>

Deletes a DataStream on a property.

pub async fn update_data_stream( &mut self, request: impl IntoRequest<UpdateDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Updates a DataStream on a property.

pub async fn list_data_streams( &mut self, request: impl IntoRequest<ListDataStreamsRequest> ) -> Result<Response<ListDataStreamsResponse>, Status>

Lists DataStreams on a property.

pub async fn get_data_stream( &mut self, request: impl IntoRequest<GetDataStreamRequest> ) -> Result<Response<DataStream>, Status>

Lookup for a single DataStream.

pub async fn get_audience( &mut self, request: impl IntoRequest<GetAudienceRequest> ) -> Result<Response<Audience>, Status>

Lookup for a single Audience. Audiences created before 2020 may not be supported. Default audiences will not show filter definitions.

pub async fn list_audiences( &mut self, request: impl IntoRequest<ListAudiencesRequest> ) -> Result<Response<ListAudiencesResponse>, Status>

Lists Audiences on a property. Audiences created before 2020 may not be supported. Default audiences will not show filter definitions.

pub async fn create_audience( &mut self, request: impl IntoRequest<CreateAudienceRequest> ) -> Result<Response<Audience>, Status>

Creates an Audience.

pub async fn update_audience( &mut self, request: impl IntoRequest<UpdateAudienceRequest> ) -> Result<Response<Audience>, Status>

Updates an Audience on a property.

pub async fn archive_audience( &mut self, request: impl IntoRequest<ArchiveAudienceRequest> ) -> Result<Response<()>, Status>

Archives an Audience on a property.

pub async fn get_search_ads360_link( &mut self, request: impl IntoRequest<GetSearchAds360LinkRequest> ) -> Result<Response<SearchAds360Link>, Status>

Look up a single SearchAds360Link

pub async fn list_search_ads360_links( &mut self, request: impl IntoRequest<ListSearchAds360LinksRequest> ) -> Result<Response<ListSearchAds360LinksResponse>, Status>

Lists all SearchAds360Links on a property.

pub async fn create_search_ads360_link( &mut self, request: impl IntoRequest<CreateSearchAds360LinkRequest> ) -> Result<Response<SearchAds360Link>, Status>

Creates a SearchAds360Link.

pub async fn delete_search_ads360_link( &mut self, request: impl IntoRequest<DeleteSearchAds360LinkRequest> ) -> Result<Response<()>, Status>

Deletes a SearchAds360Link on a property.

pub async fn update_search_ads360_link( &mut self, request: impl IntoRequest<UpdateSearchAds360LinkRequest> ) -> Result<Response<SearchAds360Link>, Status>

Updates a SearchAds360Link on a property.

pub async fn get_attribution_settings( &mut self, request: impl IntoRequest<GetAttributionSettingsRequest> ) -> Result<Response<AttributionSettings>, Status>

Lookup for a AttributionSettings singleton.

pub async fn update_attribution_settings( &mut self, request: impl IntoRequest<UpdateAttributionSettingsRequest> ) -> Result<Response<AttributionSettings>, Status>

Updates attribution settings on a property.

pub async fn run_access_report( &mut self, request: impl IntoRequest<RunAccessReportRequest> ) -> Result<Response<RunAccessReportResponse>, Status>

Returns a customized report of data access records. The report provides records of each time a user reads Google Analytics reporting data. Access records are retained for up to 2 years.

Data Access Reports can be requested for a property. Reports may be requested for any property, but dimensions that aren’t related to quota can only be requested on Google Analytics 360 properties. This method is only available to Administrators.

These data access records include GA4 UI Reporting, GA4 UI Explorations, GA4 Data API, and other products like Firebase & Admob that can retrieve data from Google Analytics through a linkage. These records don’t include property configuration changes like adding a stream or changing a property’s time zone. For configuration change history, see searchChangeHistoryEvents.

pub async fn create_access_binding( &mut self, request: impl IntoRequest<CreateAccessBindingRequest> ) -> Result<Response<AccessBinding>, Status>

Creates an access binding on an account or property.

pub async fn get_access_binding( &mut self, request: impl IntoRequest<GetAccessBindingRequest> ) -> Result<Response<AccessBinding>, Status>

Gets information about an access binding.

pub async fn update_access_binding( &mut self, request: impl IntoRequest<UpdateAccessBindingRequest> ) -> Result<Response<AccessBinding>, Status>

Updates an access binding on an account or property.

pub async fn delete_access_binding( &mut self, request: impl IntoRequest<DeleteAccessBindingRequest> ) -> Result<Response<()>, Status>

Deletes an access binding on an account or property.

pub async fn list_access_bindings( &mut self, request: impl IntoRequest<ListAccessBindingsRequest> ) -> Result<Response<ListAccessBindingsResponse>, Status>

Lists all access bindings on an account or property.

pub async fn batch_create_access_bindings( &mut self, request: impl IntoRequest<BatchCreateAccessBindingsRequest> ) -> Result<Response<BatchCreateAccessBindingsResponse>, Status>

Creates information about multiple access bindings to an account or property.

This method is transactional. If any AccessBinding cannot be created, none of the AccessBindings will be created.

pub async fn batch_get_access_bindings( &mut self, request: impl IntoRequest<BatchGetAccessBindingsRequest> ) -> Result<Response<BatchGetAccessBindingsResponse>, Status>

Gets information about multiple access bindings to an account or property.

pub async fn batch_update_access_bindings( &mut self, request: impl IntoRequest<BatchUpdateAccessBindingsRequest> ) -> Result<Response<BatchUpdateAccessBindingsResponse>, Status>

Updates information about multiple access bindings to an account or property.

pub async fn batch_delete_access_bindings( &mut self, request: impl IntoRequest<BatchDeleteAccessBindingsRequest> ) -> Result<Response<()>, Status>

Deletes information about multiple users’ links to an account or property.

pub async fn get_expanded_data_set( &mut self, request: impl IntoRequest<GetExpandedDataSetRequest> ) -> Result<Response<ExpandedDataSet>, Status>

Lookup for a single ExpandedDataSet.

pub async fn list_expanded_data_sets( &mut self, request: impl IntoRequest<ListExpandedDataSetsRequest> ) -> Result<Response<ListExpandedDataSetsResponse>, Status>

Lists ExpandedDataSets on a property.

pub async fn create_expanded_data_set( &mut self, request: impl IntoRequest<CreateExpandedDataSetRequest> ) -> Result<Response<ExpandedDataSet>, Status>

Creates a ExpandedDataSet.

pub async fn update_expanded_data_set( &mut self, request: impl IntoRequest<UpdateExpandedDataSetRequest> ) -> Result<Response<ExpandedDataSet>, Status>

Updates a ExpandedDataSet on a property.

pub async fn delete_expanded_data_set( &mut self, request: impl IntoRequest<DeleteExpandedDataSetRequest> ) -> Result<Response<()>, Status>

Deletes a ExpandedDataSet on a property.

pub async fn get_channel_group( &mut self, request: impl IntoRequest<GetChannelGroupRequest> ) -> Result<Response<ChannelGroup>, Status>

Lookup for a single ChannelGroup.

pub async fn list_channel_groups( &mut self, request: impl IntoRequest<ListChannelGroupsRequest> ) -> Result<Response<ListChannelGroupsResponse>, Status>

Lists ChannelGroups on a property.

pub async fn create_channel_group( &mut self, request: impl IntoRequest<CreateChannelGroupRequest> ) -> Result<Response<ChannelGroup>, Status>

Creates a ChannelGroup.

pub async fn update_channel_group( &mut self, request: impl IntoRequest<UpdateChannelGroupRequest> ) -> Result<Response<ChannelGroup>, Status>

Updates a ChannelGroup.

pub async fn delete_channel_group( &mut self, request: impl IntoRequest<DeleteChannelGroupRequest> ) -> Result<Response<()>, Status>

Deletes a ChannelGroup on a property.

pub async fn set_automated_ga4_configuration_opt_out( &mut self, request: impl IntoRequest<SetAutomatedGa4ConfigurationOptOutRequest> ) -> Result<Response<SetAutomatedGa4ConfigurationOptOutResponse>, Status>

Sets the opt out status for the automated GA4 setup process for a UA property. Note: this has no effect on GA4 property.

pub async fn fetch_automated_ga4_configuration_opt_out( &mut self, request: impl IntoRequest<FetchAutomatedGa4ConfigurationOptOutRequest> ) -> Result<Response<FetchAutomatedGa4ConfigurationOptOutResponse>, Status>

Fetches the opt out status for the automated GA4 setup process for a UA property. Note: this has no effect on GA4 property.

pub async fn get_big_query_link( &mut self, request: impl IntoRequest<GetBigQueryLinkRequest> ) -> Result<Response<BigQueryLink>, Status>

Lookup for a single BigQuery Link.

pub async fn list_big_query_links( &mut self, request: impl IntoRequest<ListBigQueryLinksRequest> ) -> Result<Response<ListBigQueryLinksResponse>, Status>

Lists BigQuery Links on a property.

pub async fn get_enhanced_measurement_settings( &mut self, request: impl IntoRequest<GetEnhancedMeasurementSettingsRequest> ) -> Result<Response<EnhancedMeasurementSettings>, Status>

Returns the enhanced measurement settings for this data stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn update_enhanced_measurement_settings( &mut self, request: impl IntoRequest<UpdateEnhancedMeasurementSettingsRequest> ) -> Result<Response<EnhancedMeasurementSettings>, Status>

Updates the enhanced measurement settings for this data stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn create_connected_site_tag( &mut self, request: impl IntoRequest<CreateConnectedSiteTagRequest> ) -> Result<Response<CreateConnectedSiteTagResponse>, Status>

Creates a connected site tag for a Universal Analytics property. You can create a maximum of 20 connected site tags per property. Note: This API cannot be used on GA4 properties.

pub async fn delete_connected_site_tag( &mut self, request: impl IntoRequest<DeleteConnectedSiteTagRequest> ) -> Result<Response<()>, Status>

Deletes a connected site tag for a Universal Analytics property. Note: this has no effect on GA4 properties.

pub async fn list_connected_site_tags( &mut self, request: impl IntoRequest<ListConnectedSiteTagsRequest> ) -> Result<Response<ListConnectedSiteTagsResponse>, Status>

Lists the connected site tags for a Universal Analytics property. A maximum of 20 connected site tags will be returned. Note: this has no effect on GA4 property.

pub async fn fetch_connected_ga4_property( &mut self, request: impl IntoRequest<FetchConnectedGa4PropertyRequest> ) -> Result<Response<FetchConnectedGa4PropertyResponse>, Status>

Given a specified UA property, looks up the GA4 property connected to it. Note: this cannot be used with GA4 properties.

pub async fn get_ad_sense_link( &mut self, request: impl IntoRequest<GetAdSenseLinkRequest> ) -> Result<Response<AdSenseLink>, Status>

Looks up a single AdSenseLink.

pub async fn create_ad_sense_link( &mut self, request: impl IntoRequest<CreateAdSenseLinkRequest> ) -> Result<Response<AdSenseLink>, Status>

Creates an AdSenseLink.

pub async fn delete_ad_sense_link( &mut self, request: impl IntoRequest<DeleteAdSenseLinkRequest> ) -> Result<Response<()>, Status>

Deletes an AdSenseLink.

pub async fn list_ad_sense_links( &mut self, request: impl IntoRequest<ListAdSenseLinksRequest> ) -> Result<Response<ListAdSenseLinksResponse>, Status>

Lists AdSenseLinks on a property.

pub async fn get_event_create_rule( &mut self, request: impl IntoRequest<GetEventCreateRuleRequest> ) -> Result<Response<EventCreateRule>, Status>

Lookup for a single EventCreateRule.

pub async fn list_event_create_rules( &mut self, request: impl IntoRequest<ListEventCreateRulesRequest> ) -> Result<Response<ListEventCreateRulesResponse>, Status>

Lists EventCreateRules on a web data stream.

pub async fn create_event_create_rule( &mut self, request: impl IntoRequest<CreateEventCreateRuleRequest> ) -> Result<Response<EventCreateRule>, Status>

Creates an EventCreateRule.

pub async fn update_event_create_rule( &mut self, request: impl IntoRequest<UpdateEventCreateRuleRequest> ) -> Result<Response<EventCreateRule>, Status>

Updates an EventCreateRule.

pub async fn delete_event_create_rule( &mut self, request: impl IntoRequest<DeleteEventCreateRuleRequest> ) -> Result<Response<()>, Status>

Deletes an EventCreateRule.

pub async fn update_data_redaction_settings( &mut self, request: impl IntoRequest<UpdateDataRedactionSettingsRequest> ) -> Result<Response<DataRedactionSettings>, Status>

Updates a DataRedactionSettings on a property.

pub async fn get_data_redaction_settings( &mut self, request: impl IntoRequest<GetDataRedactionSettingsRequest> ) -> Result<Response<DataRedactionSettings>, Status>

Lookup for a single DataRedactionSettings.

pub async fn get_calculated_metric( &mut self, request: impl IntoRequest<GetCalculatedMetricRequest> ) -> Result<Response<CalculatedMetric>, Status>

Lookup for a single CalculatedMetric.

pub async fn create_calculated_metric( &mut self, request: impl IntoRequest<CreateCalculatedMetricRequest> ) -> Result<Response<CalculatedMetric>, Status>

Creates a CalculatedMetric.

pub async fn list_calculated_metrics( &mut self, request: impl IntoRequest<ListCalculatedMetricsRequest> ) -> Result<Response<ListCalculatedMetricsResponse>, Status>

Lists CalculatedMetrics on a property.

pub async fn update_calculated_metric( &mut self, request: impl IntoRequest<UpdateCalculatedMetricRequest> ) -> Result<Response<CalculatedMetric>, Status>

Updates a CalculatedMetric on a property.

pub async fn delete_calculated_metric( &mut self, request: impl IntoRequest<DeleteCalculatedMetricRequest> ) -> Result<Response<()>, Status>

Deletes a CalculatedMetric on a property.

pub async fn create_rollup_property( &mut self, request: impl IntoRequest<CreateRollupPropertyRequest> ) -> Result<Response<CreateRollupPropertyResponse>, Status>

Create a roll-up property and all roll-up property source links.

pub async fn get_rollup_property_source_link( &mut self, request: impl IntoRequest<GetRollupPropertySourceLinkRequest> ) -> Result<Response<RollupPropertySourceLink>, Status>

Lookup for a single roll-up property source Link. Only roll-up properties can have source links, so this method will throw an error if used on other types of properties.

pub async fn list_rollup_property_source_links( &mut self, request: impl IntoRequest<ListRollupPropertySourceLinksRequest> ) -> Result<Response<ListRollupPropertySourceLinksResponse>, Status>

Lists roll-up property source Links on a property. Only roll-up properties can have source links, so this method will throw an error if used on other types of properties.

pub async fn create_rollup_property_source_link( &mut self, request: impl IntoRequest<CreateRollupPropertySourceLinkRequest> ) -> Result<Response<RollupPropertySourceLink>, Status>

Creates a roll-up property source link. Only roll-up properties can have source links, so this method will throw an error if used on other types of properties.

pub async fn delete_rollup_property_source_link( &mut self, request: impl IntoRequest<DeleteRollupPropertySourceLinkRequest> ) -> Result<Response<()>, Status>

Deletes a roll-up property source link. Only roll-up properties can have source links, so this method will throw an error if used on other types of properties.

pub async fn create_subproperty( &mut self, request: impl IntoRequest<CreateSubpropertyRequest> ) -> Result<Response<CreateSubpropertyResponse>, Status>

Create a subproperty and a subproperty event filter that applies to the created subproperty.

pub async fn create_subproperty_event_filter( &mut self, request: impl IntoRequest<CreateSubpropertyEventFilterRequest> ) -> Result<Response<SubpropertyEventFilter>, Status>

Creates a subproperty Event Filter.

pub async fn get_subproperty_event_filter( &mut self, request: impl IntoRequest<GetSubpropertyEventFilterRequest> ) -> Result<Response<SubpropertyEventFilter>, Status>

Lookup for a single subproperty Event Filter.

pub async fn list_subproperty_event_filters( &mut self, request: impl IntoRequest<ListSubpropertyEventFiltersRequest> ) -> Result<Response<ListSubpropertyEventFiltersResponse>, Status>

List all subproperty Event Filters on a property.

pub async fn update_subproperty_event_filter( &mut self, request: impl IntoRequest<UpdateSubpropertyEventFilterRequest> ) -> Result<Response<SubpropertyEventFilter>, Status>

Updates a subproperty Event Filter.

pub async fn delete_subproperty_event_filter( &mut self, request: impl IntoRequest<DeleteSubpropertyEventFilterRequest> ) -> Result<Response<()>, Status>

Deletes a subproperty event filter.