Module google_api_proto::google::analytics::data::v1alpha

Modules

alpha_analytics_data_client

Generated client implementations.

audience_dimension_value

Nested message and enum types in AudienceDimensionValue.

audience_list

Nested message and enum types in AudienceList.

dimension_expression

Nested message and enum types in DimensionExpression.

dimension_value

Nested message and enum types in DimensionValue.

filter

Nested message and enum types in Filter.

filter_expression

Nested message and enum types in FilterExpression.

funnel_field_filter

Nested message and enum types in FunnelFieldFilter.

funnel_filter_expression

Nested message and enum types in FunnelFilterExpression.

funnel_parameter_filter

Nested message and enum types in FunnelParameterFilter.

funnel_parameter_filter_expression

Nested message and enum types in FunnelParameterFilterExpression.

metric_value

Nested message and enum types in MetricValue.

numeric_filter

Nested message and enum types in NumericFilter.

numeric_value

Nested message and enum types in NumericValue.

run_funnel_report_request

Nested message and enum types in RunFunnelReportRequest.

segment

Nested message and enum types in Segment.

segment_filter

Nested message and enum types in SegmentFilter.

segment_filter_expression

Nested message and enum types in SegmentFilterExpression.

segment_parameter_filter

Nested message and enum types in SegmentParameterFilter.

segment_parameter_filter_expression

Nested message and enum types in SegmentParameterFilterExpression.

string_filter

Nested message and enum types in StringFilter.

Structs

AudienceDimension

An audience dimension is a user attribute. Specific user attributed are requested and then later returned in the QueryAudienceListResponse.

AudienceDimensionValue

The value of a dimension.

AudienceList

An audience list is a list of users in an audience at the time of the list’s creation. One audience may have multiple audience lists created for different days.

AudienceListMetadata

This metadata is currently blank.

AudienceRow

Dimension value attributes for the audience user row.

BetweenFilter

To express that the result needs to be between two numbers (inclusive).

CreateAudienceListRequest

A request to create a new audience list.

CreateRecurringAudienceListRequest

A request to create a new recurring audience list.

DateRange

A contiguous set of days: startDate, startDate + 1, …, endDate. Requests are allowed up to 4 date ranges.

Dimension

Dimensions are attributes of your data. For example, the dimension city indicates the city from which an event originates. Dimension values in report responses are strings; for example, the city could be “Paris” or “New York”.

DimensionExpression

Used to express a dimension which is the result of a formula of multiple dimensions. Example usages:

DimensionHeader

Describes a dimension column in the report. Dimensions requested in a report produce column entries within rows and DimensionHeaders. However, dimensions used exclusively within filters or expressions do not produce columns in a report; correspondingly, those dimensions do not produce headers.

DimensionValue

The value of a dimension.

EventSegment

Event segments are subsets of events that were triggered on your site or app. for example, all purchase events made in a particular location; app_exception events that occurred on a specific operating system.

EventSegmentConditionGroup

Conditions tell Analytics what data to include in or exclude from the segment.

EventSegmentCriteria

An event matches a criteria if the event meet the conditions in the criteria.

EventSegmentExclusion

Specifies which events are excluded in this segment.

Filter

An expression to filter dimension or metric values.

FilterExpression

To express dimension or metric filters. The fields in the same FilterExpression need to be either all dimensions or all metrics.

FilterExpressionList

A list of filter expressions.

Funnel

Configures the funnel in a funnel report request. A funnel reports on users as they pass through a sequence of steps.

FunnelBreakdown

Breakdowns add a dimension to the funnel table sub report response.

FunnelEventFilter

Creates a filter that matches events of a single event name. If a parameter filter expression is specified, only the subset of events that match both the single event name and the parameter filter expressions match this event filter.

FunnelFieldFilter

An expression to filter dimension or metric values.

FunnelFilterExpression

Expresses combinations of funnel filters.

FunnelFilterExpressionList

A list of funnel filter expressions.

FunnelNextAction

Next actions state the value for a dimension after the user has achieved a step but before the same user has achieved the next step. For example if the nextActionDimension is eventName, then nextActionDimension in the ith funnel step row will return first event after the event that qualified the user into the ith funnel step but before the user achieved the i+1th funnel step.

FunnelParameterFilter

An expression to filter parameter values in a funnel.

FunnelParameterFilterExpression

Expresses combinations of funnel filters on parameters.

FunnelParameterFilterExpressionList

A list of funnel parameter filter expressions.

FunnelResponseMetadata

The funnel report’s response metadata carries additional information about the funnel report.

FunnelStep

Steps define the user journey you want to measure. Steps contain one or more conditions that your users must meet to be included in that step of the funnel journey.

FunnelSubReport

Funnel sub reports contain the dimension and metric data values. For example, 12 users reached the second step of the funnel.

GetAudienceListRequest

A request to retrieve configuration metadata about a specific audience list.

GetRecurringAudienceListRequest

A request to retrieve configuration metadata about a specific recurring audience list.

InListFilter

The result needs to be in a list of string values.

ListAudienceListsRequest

A request to list all audience lists for a property.

ListAudienceListsResponse

A list of all audience lists for a property.

ListRecurringAudienceListsRequest

A request to list all recurring audience lists for a property.

ListRecurringAudienceListsResponse

A list of all recurring audience lists for a property.

MetricHeader

Describes a metric column in the report. Visible metrics requested in a report produce column entries within rows and MetricHeaders. However, metrics used exclusively within filters or expressions do not produce columns in a report; correspondingly, those metrics do not produce headers.

MetricValue

The value of a metric.

NumericFilter

Filters for numeric or date values.

NumericValue

To represent a number.

PropertyQuota

Current state of all quotas for this Analytics Property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

QueryAudienceListRequest

A request to list users in an audience list.

QueryAudienceListResponse

A list of users in an audience list.

QuotaStatus

Current state for a particular quota group.

RecurringAudienceList

A recurring audience list produces new audience lists each day. Audience lists are users in an audience at the time of the list’s creation. A recurring audience list ensures that you have audience list based on the most recent data available for use each day.

Row

Report data for each row. For example if RunReportRequest contains:

RunFunnelReportRequest

The request for a funnel report.

RunFunnelReportResponse

The funnel report response contains two sub reports. The two sub reports are different combinations of dimensions and metrics.

SamplingMetadata

If funnel report results are sampled, this metadata describes what percentage of events were used in this funnel report for a date range. Sampling is the practice of analyzing a subset of all data in order to uncover the meaningful information in the larger data set.

Segment

A segment is a subset of your Analytics data. For example, of your entire set of users, one segment might be users from a particular country or city. Another segment might be users who purchase a particular line of products or who visit a specific part of your site or trigger certain events in your app.

SegmentEventFilter

Creates a filter that matches events of a single event name. If a parameter filter expression is specified, only the subset of events that match both the single event name and the parameter filter expressions match this event filter.

SegmentFilter

An expression to filter dimension or metric values.

SegmentFilterExpression

Expresses combinations of segment filters.

SegmentFilterExpressionList

A list of segment filter expressions.

SegmentFilterScoping

Scopings specify how the dimensions & metrics of multiple events should be considered when evaluating a segment filter.

SegmentParameterFilter

An expression to filter parameter values in a segment.

SegmentParameterFilterExpression

Expresses combinations of segment filter on parameters.

SegmentParameterFilterExpressionList

A list of segment parameter filter expressions.

SegmentParameterFilterScoping

Scopings specify how multiple events should be considered when evaluating a segment parameter filter.

SessionSegment

Session segments are subsets of the sessions that occurred on your site or app: for example, all the sessions that originated from a particular advertising campaign.

SessionSegmentConditionGroup

Conditions tell Analytics what data to include in or exclude from the segment.

SessionSegmentCriteria

A session matches a criteria if the session’s events meet the conditions in the criteria.

SessionSegmentExclusion

Specifies which sessions are excluded in this segment.

SheetExportAudienceListRequest

A request to export users in an audience list to a Google Sheet.

SheetExportAudienceListResponse

The created Google Sheet with the list of users in an audience list.

StringFilter

The filter for string

UserSegment

User segments are subsets of users who engaged with your site or app. For example, users who have previously purchased; users who added items to their shopping carts, but didn’t complete a purchase.

UserSegmentConditionGroup

Conditions tell Analytics what data to include in or exclude from the segment.

UserSegmentCriteria

A user matches a criteria if the user’s events meet the conditions in the criteria.

UserSegmentExclusion

Specifies which users are excluded in this segment.

UserSegmentSequenceGroup

Define conditions that must occur in a specific order for the user to be a member of the segment.

UserSequenceStep

A condition that must occur in the specified step order for this user to match the sequence.

WebhookNotification

Configures a long-running operation resource to send a webhook notification from the Google Analytics Data API to your webhook server when the resource updates.

Enums

EventCriteriaScoping

Scoping specifies which events are considered when evaluating if an event meets a criteria.

EventExclusionDuration

Enumerates options for how long an exclusion will last if an event matches the eventExclusionCriteria.

MetricType

A metric’s value type.

SessionCriteriaScoping

Scoping specifies which events are considered when evaluating if a session meets a criteria.

SessionExclusionDuration

Enumerates options for how long an exclusion will last if a session matches the sessionExclusionCriteria.

UserCriteriaScoping

Scoping specifies which events are considered when evaluating if a user meets a criteria.

UserExclusionDuration

Enumerates options for how long an exclusion will last if a user matches the userExclusionCriteria.