Module google_api_proto::google::api::servicemanagement::v1

Modules

config_file

Nested message and enum types in ConfigFile.

diagnostic

Nested message and enum types in Diagnostic.

get_service_config_request

Nested message and enum types in GetServiceConfigRequest.

operation_metadata

Nested message and enum types in OperationMetadata.

rollout

Nested message and enum types in Rollout.

service_manager_client

Generated client implementations.

Structs

ChangeReport

Change report associated with a particular service configuration.

ConfigFile

Generic specification of a source configuration file

ConfigRef

Represents a service configuration with its name and id.

ConfigSource

Represents a source file which is used to generate the service configuration defined by google.api.Service.

CreateServiceConfigRequest

Request message for CreateServiceConfig method.

CreateServiceRequest

Request message for CreateService method.

CreateServiceRolloutRequest

Request message for ‘CreateServiceRollout’

DeleteServiceRequest

Request message for DeleteService method.

Diagnostic

Represents a diagnostic message (error or warning)

GenerateConfigReportRequest

Request message for GenerateConfigReport method.

GenerateConfigReportResponse

Response message for GenerateConfigReport method.

GetServiceConfigRequest

Request message for GetServiceConfig method.

GetServiceRequest

Request message for GetService method.

GetServiceRolloutRequest

Request message for GetServiceRollout method.

ListServiceConfigsRequest

Request message for ListServiceConfigs method.

ListServiceConfigsResponse

Response message for ListServiceConfigs method.

ListServiceRolloutsRequest

Request message for ‘ListServiceRollouts’

ListServiceRolloutsResponse

Response message for ListServiceRollouts method.

ListServicesRequest

Request message for ListServices method.

ListServicesResponse

Response message for ListServices method.

ManagedService

The full representation of a Service that is managed by Google Service Management.

OperationMetadata

The metadata associated with a long running operation resource.

Rollout

A rollout resource that defines how service configuration versions are pushed to control plane systems. Typically, you create a new version of the service config, and then create a Rollout to push the service config.

SubmitConfigSourceRequest

Request message for SubmitConfigSource method.

SubmitConfigSourceResponse

Response message for SubmitConfigSource method.

UndeleteServiceRequest

Request message for UndeleteService method.

UndeleteServiceResponse

Response message for UndeleteService method.