Struct google_api_proto::google::appengine::v1beta::Version

pub struct Version {

    pub name: String,
    pub id: String,
    pub inbound_services: Vec<i32>,
    pub instance_class: String,
    pub network: Option<Network>,
    pub zones: Vec<String>,
    pub resources: Option<Resources>,
    pub runtime: String,
    pub runtime_channel: String,
    pub threadsafe: bool,
    pub vm: bool,
    pub app_engine_apis: bool,
    pub beta_settings: BTreeMap<String, String>,
    pub env: String,
    pub serving_status: i32,
    pub created_by: String,
    pub create_time: Option<Timestamp>,
    pub disk_usage_bytes: i64,
    pub runtime_api_version: String,
    pub runtime_main_executable_path: String,
    pub service_account: String,
    pub handlers: Vec<UrlMap>,
    pub error_handlers: Vec<ErrorHandler>,
    pub libraries: Vec<Library>,
    pub api_config: Option<ApiConfigHandler>,
    pub env_variables: BTreeMap<String, String>,
    pub build_env_variables: BTreeMap<String, String>,
    pub default_expiration: Option<Duration>,
    pub health_check: Option<HealthCheck>,
    pub readiness_check: Option<ReadinessCheck>,
    pub liveness_check: Option<LivenessCheck>,
    pub nobuild_files_regex: String,
    pub deployment: Option<Deployment>,
    pub version_url: String,
    pub endpoints_api_service: Option<EndpointsApiService>,
    pub entrypoint: Option<Entrypoint>,
    pub vpc_access_connector: Option<VpcAccessConnector>,
    pub scaling: Option<Scaling>,

}

A Version resource is a specific set of source code and configuration files that are deployed into a service.

Fields
name: String

Full path to the Version resource in the API. Example: apps/myapp/services/default/versions/v1.

@OutputOnly

id: String

Relative name of the version within the service. Example: v1. Version names can contain only lowercase letters, numbers, or hyphens. Reserved names: “default”, “latest”, and any name with the prefix “ah-”.

inbound_services: Vec<i32>

Before an application can receive email or XMPP messages, the application must be configured to enable the service.

instance_class: String

Instance class that is used to run this version. Valid values are:

AutomaticScaling: F1, F2, F4, F4_1G

ManualScaling or BasicScaling: B1, B2, B4, B8, B4_1G

Defaults to F1 for AutomaticScaling and B1 for ManualScaling or BasicScaling.

network: Option<Network>

Extra network settings. Only applicable in the App Engine flexible environment.

zones: Vec<String>

The Google Compute Engine zones that are supported by this version in the App Engine flexible environment. Deprecated.

resources: Option<Resources>

Machine resources for this version. Only applicable in the App Engine flexible environment.

runtime: String

Desired runtime. Example: python27.

runtime_channel: String

The channel of the runtime to use. Only available for some runtimes. Defaults to the default channel.

threadsafe: bool

Whether multiple requests can be dispatched to this version at once.

vm: bool

Whether to deploy this version in a container on a virtual machine.

app_engine_apis: bool

app_engine_apis allows second generation runtimes to access the App Engine APIs.

beta_settings: BTreeMap<String, String>

Metadata settings that are supplied to this version to enable beta runtime features.

env: String

App Engine execution environment for this version.

Defaults to standard.

serving_status: i32

Current serving status of this version. Only the versions with a SERVING status create instances and can be billed.

SERVING_STATUS_UNSPECIFIED is an invalid value. Defaults to SERVING.

created_by: String

Email address of the user who created this version.

@OutputOnly

create_time: Option<Timestamp>

Time that this version was created.

@OutputOnly

disk_usage_bytes: i64

Total size in bytes of all the files that are included in this version and currently hosted on the App Engine disk.

@OutputOnly

runtime_api_version: String

The version of the API in the given runtime environment. Please see the app.yaml reference for valid values at <https://cloud.google.com/appengine/docs/standard//config/appref>

runtime_main_executable_path: String

The path or name of the app’s main executable.

service_account: String

The identity that the deployed version will run as. Admin API will use the App Engine Appspot service account as default if this field is neither provided in app.yaml file nor through CLI flag.

handlers: Vec<UrlMap>

An ordered list of URL-matching patterns that should be applied to incoming requests. The first matching URL handles the request and other request handlers are not attempted.

Only returned in GET requests if view=FULL is set.

error_handlers: Vec<ErrorHandler>

Custom static error pages. Limited to 10KB per page.

Only returned in GET requests if view=FULL is set.

libraries: Vec<Library>

Configuration for third-party Python runtime libraries that are required by the application.

Only returned in GET requests if view=FULL is set.

api_config: Option<ApiConfigHandler>

Serving configuration for Google Cloud Endpoints.

Only returned in GET requests if view=FULL is set.

env_variables: BTreeMap<String, String>

Environment variables available to the application.

Only returned in GET requests if view=FULL is set.

build_env_variables: BTreeMap<String, String>

Environment variables available to the build environment.

Only returned in GET requests if view=FULL is set.

default_expiration: Option<Duration>

Duration that static files should be cached by web proxies and browsers. Only applicable if the corresponding StaticFilesHandler does not specify its own expiration time.

Only returned in GET requests if view=FULL is set.

health_check: Option<HealthCheck>

Configures health checking for instances. Unhealthy instances are stopped and replaced with new instances. Only applicable in the App Engine flexible environment.

Only returned in GET requests if view=FULL is set.

readiness_check: Option<ReadinessCheck>

Configures readiness health checking for instances. Unhealthy instances are not put into the backend traffic rotation.

Only returned in GET requests if view=FULL is set.

liveness_check: Option<LivenessCheck>

Configures liveness health checking for instances. Unhealthy instances are stopped and replaced with new instances

Only returned in GET requests if view=FULL is set.

nobuild_files_regex: String

Files that match this pattern will not be built into this version. Only applicable for Go runtimes.

Only returned in GET requests if view=FULL is set.

deployment: Option<Deployment>

Code and application artifacts that make up this version.

Only returned in GET requests if view=FULL is set.

version_url: String

Serving URL for this version. Example: “https://myversion-dot-myservice-dot-myapp.appspot.com"

@OutputOnly

endpoints_api_service: Option<EndpointsApiService>

Cloud Endpoints configuration.

If endpoints_api_service is set, the Cloud Endpoints Extensible Service Proxy will be provided to serve the API implemented by the app.

entrypoint: Option<Entrypoint>

The entrypoint for the application.

vpc_access_connector: Option<VpcAccessConnector>

Enables VPC connectivity for standard apps.

scaling: Option<Scaling>

Controls how instances are created, scaled, and reaped.

Defaults to AutomaticScaling.

Implementations

impl Version

pub fn inbound_services( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<InboundServiceType>>

Returns an iterator which yields the valid enum values contained in inbound_services.

pub fn push_inbound_services(&mut self, value: InboundServiceType)

Appends the provided enum value to inbound_services.

pub fn serving_status(&self) -> ServingStatus

Returns the enum value of serving_status, or the default if the field is set to an invalid enum value.

pub fn set_serving_status(&mut self, value: ServingStatus)

Sets serving_status to the provided enum value.