Module google_api_proto::google::apps::card::v1

Modules

action

Nested message and enum types in Action.

border_style

Nested message and enum types in BorderStyle.

card

Nested message and enum types in Card.

columns

Nested message and enum types in Columns.

date_time_picker

Nested message and enum types in DateTimePicker.

decorated_text

Nested message and enum types in DecoratedText.

grid

Nested message and enum types in Grid.

icon

Nested message and enum types in Icon.

image_crop_style

Nested message and enum types in ImageCropStyle.

on_click

Nested message and enum types in OnClick.

open_link

Nested message and enum types in OpenLink.

selection_input

Nested message and enum types in SelectionInput.

suggestions

Nested message and enum types in Suggestions.

text_input

Nested message and enum types in TextInput.

widget

Nested message and enum types in Widget.

Structs

Action

An action that describes the behavior when the form is submitted. For example, you can invoke an Apps Script script to handle the form. If the action is triggered, the form values are sent to the server.

BorderStyle

The style options for the border of a card or widget, including the border type and color.

Button

A text, icon, or text and icon button that users can click. For an example in Google Chat apps, see Button list.

ButtonList

A list of buttons layed out horizontally. For an example in Google Chat apps, see Button list.

Card

A card interface displayed in a Google Chat message or Google Workspace Add-on.

Columns

The Columns widget displays up to 2 columns in a card message or dialog. You can add widgets to each column; the widgets appear in the order that they are specified. For an example in Google Chat apps, see Columns.

DateTimePicker

Lets users input a date, a time, or both a date and a time. For an example in Google Chat apps, see Date time picker.

DecoratedText

A widget that displays text with optional decorations such as a label above or below the text, an icon in front of the text, a selection widget, or a button after the text. For an example in Google Chat apps, see Decorated text.

Divider

Displays a divider between widgets as a horizontal line. For an example in Google Chat apps, see Divider.

Grid

Displays a grid with a collection of items. Items can only include text or images. For responsive columns, or to include more than text or images, use [Columns][google.apps.card.v1.Columns]. For an example in Google Chat apps, see Grid.

Icon

An icon displayed in a widget on a card. For an example in Google Chat apps, see Icon.

Image

An image that is specified by a URL and can have an onClick action. For an example, see Image.

ImageComponent

Represents an image.

ImageCropStyle

Represents the crop style applied to an image.

OnClick

Represents how to respond when users click an interactive element on a card, such as a button.

OpenLink

Represents an onClick event that opens a hyperlink.

SelectionInput

A widget that creates one or more UI items that users can select. For example, a dropdown menu or checkboxes. You can use this widget to collect data that can be predicted or enumerated. For an example in Google Chat apps, see Selection input.

Suggestions

Suggested values that users can enter. These values appear when users click inside the text input field. As users type, the suggested values dynamically filter to match what the users have typed.

TextInput

A field in which users can enter text. Supports suggestions and on-change actions. For an example in Google Chat apps, see Text input.

TextParagraph

A paragraph of text that supports formatting. For an example in Google Chat apps, see Text paragraph. For more information about formatting text, see Formatting text in Google Chat apps and Formatting text in Google Workspace Add-ons.

Widget

Each card is made up of widgets.