Module google_api_proto::google::apps::script::type

Modules

add_on_widget_set

Nested message and enum types in AddOnWidgetSet.

calendar

docs

drive

gmail

sheets

slides

universal_action_extension_point

Nested message and enum types in UniversalActionExtensionPoint.

Structs

AddOnWidgetSet

The widget subset used by an add-on.

CommonAddOnManifest

Add-on configuration that is shared across all add-on host applications.

HomepageExtensionPoint

Common format for declaring an add-on’s home-page view.

HttpOptions

Options for sending requests to add-on HTTP endpoints

LayoutProperties

Card layout properties shared across all add-on host applications.

MenuItemExtensionPoint

Common format for declaring a menu item, or button, that appears within a host app.

UniversalActionExtensionPoint

Format for declaring a universal action menu item extension point.

Enums

HttpAuthorizationHeader

Authorization header sent in add-on HTTP requests