Module google_api_proto::google::bigtable::v2

Modules

bigtable_client

Generated client implementations.

column_range

Nested message and enum types in ColumnRange.

mutate_rows_request

Nested message and enum types in MutateRowsRequest.

mutate_rows_response

Nested message and enum types in MutateRowsResponse.

mutation

Nested message and enum types in Mutation.

read_modify_write_rule

Nested message and enum types in ReadModifyWriteRule.

read_rows_response

Nested message and enum types in ReadRowsResponse.

row_filter

Nested message and enum types in RowFilter.

row_range

Nested message and enum types in RowRange.

value_range

Nested message and enum types in ValueRange.

Structs

Cell

Specifies (some of) the contents of a single row/column/timestamp of a table.

CheckAndMutateRowRequest

Request message for Bigtable.CheckAndMutateRow.

CheckAndMutateRowResponse

Response message for Bigtable.CheckAndMutateRow.

Column

Specifies (some of) the contents of a single row/column intersection of a table.

ColumnRange

Specifies a contiguous range of columns within a single column family. The range spans from <column_family>:<start_qualifier> to <column_family>:<end_qualifier>, where both bounds can be either inclusive or exclusive.

Family

Specifies (some of) the contents of a single row/column family intersection of a table.

MutateRowRequest

Request message for Bigtable.MutateRow.

MutateRowResponse

Response message for Bigtable.MutateRow.

MutateRowsRequest

Request message for BigtableService.MutateRows.

MutateRowsResponse

Response message for BigtableService.MutateRows.

Mutation

Specifies a particular change to be made to the contents of a row.

PingAndWarmRequest

Request message for client connection keep-alive and warming.

PingAndWarmResponse

Response message for Bigtable.PingAndWarm connection keepalive and warming.

ReadModifyWriteRowRequest

Request message for Bigtable.ReadModifyWriteRow.

ReadModifyWriteRowResponse

Response message for Bigtable.ReadModifyWriteRow.

ReadModifyWriteRule

Specifies an atomic read/modify/write operation on the latest value of the specified column.

ReadRowsRequest

Request message for Bigtable.ReadRows.

ReadRowsResponse

Response message for Bigtable.ReadRows.

Row

Specifies the complete (requested) contents of a single row of a table. Rows which exceed 256MiB in size cannot be read in full.

RowFilter

Takes a row as input and produces an alternate view of the row based on specified rules. For example, a RowFilter might trim down a row to include just the cells from columns matching a given regular expression, or might return all the cells of a row but not their values. More complicated filters can be composed out of these components to express requests such as, “within every column of a particular family, give just the two most recent cells which are older than timestamp X.”

RowRange

Specifies a contiguous range of rows.

RowSet

Specifies a non-contiguous set of rows.

SampleRowKeysRequest

Request message for Bigtable.SampleRowKeys.

SampleRowKeysResponse

Response message for Bigtable.SampleRowKeys.

TimestampRange

Specified a contiguous range of microsecond timestamps.

ValueRange

Specifies a contiguous range of raw byte values.