Module google_api_proto::google::chat::v1

Modules

action_response

Nested message and enum types in ActionResponse.

annotation

Nested message and enum types in Annotation.

attachment

Nested message and enum types in Attachment.

chat_service_client

Generated client implementations.

contextual_add_on_markup

Nested message and enum types in ContextualAddOnMarkup.

create_message_request

Nested message and enum types in CreateMessageRequest.

deletion_metadata

Nested message and enum types in DeletionMetadata.

dialog_action

Nested message and enum types in DialogAction.

emoji

Nested message and enum types in Emoji.

membership

Nested message and enum types in Membership.

slash_command_metadata

Nested message and enum types in SlashCommandMetadata.

space

Nested message and enum types in Space.

user

Nested message and enum types in User.

user_mention_metadata

Nested message and enum types in UserMentionMetadata.

widget_markup

Nested message and enum types in WidgetMarkup.

Structs

ActionResponse

Parameters that a Chat app can use to configure how its response is posted.

ActionStatus

Represents the status for a request to either invoke or submit a dialog.

Annotation

Output only. Annotations associated with the plain-text body of the message. To add basic formatting to a text message, see Format text messages.

AttachedGif

A GIF image that’s specified by a URL.

Attachment

An attachment in Google Chat.

AttachmentDataRef

CardWithId

A card in a Google Chat message.

CompleteImportSpaceRequest

Request message for completing the import process for a space.

CompleteImportSpaceResponse

ContextualAddOnMarkup

The markup for developers to specify the contents of a contextual AddOn.

CreateMembershipRequest

CreateMessageRequest

Creates a message.

CreateReactionRequest

Creates a reaction to a message.

CreateSpaceRequest

CustomEmoji

Represents a custom emoji.

DeleteMembershipRequest

DeleteMessageRequest

DeleteReactionRequest

Deletes a reaction to a message.

DeleteSpaceRequest

Request for deleting a space.

DeletionMetadata

Information about a deleted message. A message is deleted when delete_time is set.

Dialog

Wrapper around the card body of the dialog.

DialogAction

Contains a dialog and request status code.

DriveDataRef

A reference to the data of a drive attachment.

Emoji

An emoji that is used as a reaction to a message.

EmojiReactionSummary

The number of people who reacted to a message with a specific emoji.

FindDirectMessageRequest

A request to get direct message space based on the user resource.

GetAttachmentRequest

GetMembershipRequest

GetMessageRequest

GetSpaceRequest

A request to return a single space.

Group

A Google Group in Google Chat.

ListMembershipsRequest

ListMembershipsResponse

ListMessagesRequest

Lists messages in the specified space, that the user is a member of.

ListMessagesResponse

ListReactionsRequest

Lists reactions to a message.

ListReactionsResponse

ListSpacesRequest

A request to list the spaces the caller is a member of.

ListSpacesResponse

MatchedUrl

A matched URL in a Chat message. Chat apps can preview matched URLs. For more information, see Preview links.

Membership

Represents a membership relation in Google Chat, such as whether a user or Chat app is invited to, part of, or absent from a space.

Message

A message in a Google Chat space.

QuotedMessageMetadata

Information about a quoted message.

Reaction

A reaction to a message.

SetUpSpaceRequest

SlashCommand

A slash command in Google Chat.

SlashCommandMetadata

Annotation metadata for slash commands (/).

Space

A space in Google Chat. Spaces are conversations between two or more users or 1:1 messages between a user and a Chat app.

Thread

A thread in a Google Chat space. For example usage, see Start or reply to a message thread.

UpdateMessageRequest

UpdateSpaceRequest

A request to update a single space.

UploadAttachmentRequest

UploadAttachmentResponse

User

A user in Google Chat. When returned as an output from a request, if your Chat app authenticates as a user, the output for a User resource only populates the user’s name and type.

UserMentionMetadata

Annotation metadata for user mentions (@).

WidgetMarkup

A widget is a UI element that presents text and images.

Enums

AnnotationType

Type of the annotation.

HistoryState

The history state for messages and spaces. Specifies how long messages and conversation threads are kept after creation.