Struct google_api_proto::google::chat::v1::chat_service_client::ChatServiceClient

pub struct ChatServiceClient<T> { /* private fields */ }

Enables developers to build Chat apps and integrations on Google Chat Platform.

Implementations

impl<T> ChatServiceClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> ChatServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn create_message( &mut self, request: impl IntoRequest<CreateMessageRequest> ) -> Result<Response<Message>, Status>

Creates a message in a Google Chat space. For an example, see Create a message.

Calling this method requires authentication and supports the following authentication types:

For text messages, user authentication or app authentication are supported.

For card messages, only app authentication is supported. (Only Chat apps can create card messages.)

pub async fn list_messages( &mut self, request: impl IntoRequest<ListMessagesRequest> ) -> Result<Response<ListMessagesResponse>, Status>

Lists messages in a space that the caller is a member of, including messages from blocked members and spaces. For an example, see List messages. Requires user authentication.

pub async fn list_memberships( &mut self, request: impl IntoRequest<ListMembershipsRequest> ) -> Result<Response<ListMembershipsResponse>, Status>

Lists memberships in a space. For an example, see List memberships. Listing memberships with app authentication lists memberships in spaces that the Chat app has access to, but excludes Chat app memberships, including its own. Listing memberships with User authentication lists memberships in spaces that the authenticated user has access to.

Requires authentication. Supports app authentication and user authentication.

pub async fn get_membership( &mut self, request: impl IntoRequest<GetMembershipRequest> ) -> Result<Response<Membership>, Status>

Returns details about a membership. For an example, see Get a membership.

Requires authentication. Supports app authentication and user authentication.

pub async fn get_message( &mut self, request: impl IntoRequest<GetMessageRequest> ) -> Result<Response<Message>, Status>

Returns details about a message. For an example, see Read a message.

Requires authentication. Supports app authentication and user authentication.

Note: Might return a message from a blocked member or space.

pub async fn update_message( &mut self, request: impl IntoRequest<UpdateMessageRequest> ) -> Result<Response<Message>, Status>

Updates a message. There’s a difference between the patch and update methods. The patch method uses a patch request while the update method uses a put request. We recommend using the patch method. For an example, see Update a message.

Requires authentication. Supports app authentication and user authentication. When using app authentication, requests can only update messages created by the calling Chat app.

pub async fn delete_message( &mut self, request: impl IntoRequest<DeleteMessageRequest> ) -> Result<Response<()>, Status>

Deletes a message. For an example, see Delete a message.

Requires authentication. Supports app authentication and user authentication. When using app authentication, requests can only delete messages created by the calling Chat app.

pub async fn get_attachment( &mut self, request: impl IntoRequest<GetAttachmentRequest> ) -> Result<Response<Attachment>, Status>

Gets the metadata of a message attachment. The attachment data is fetched using the media API. For an example, see Get a message attachment. Requires app authentication.

pub async fn upload_attachment( &mut self, request: impl IntoRequest<UploadAttachmentRequest> ) -> Result<Response<UploadAttachmentResponse>, Status>

Uploads an attachment. For an example, see Upload media as a file attachment. Requires user authentication.

You can upload attachments up to 200 MB. Certain file types aren’t supported. For details, see File types blocked by Google Chat.

pub async fn list_spaces( &mut self, request: impl IntoRequest<ListSpacesRequest> ) -> Result<Response<ListSpacesResponse>, Status>

Lists spaces the caller is a member of. Group chats and DMs aren’t listed until the first message is sent. For an example, see List spaces.

Requires authentication. Supports app authentication and user authentication.

Lists spaces visible to the caller or authenticated user. Group chats and DMs aren’t listed until the first message is sent.

pub async fn get_space( &mut self, request: impl IntoRequest<GetSpaceRequest> ) -> Result<Response<Space>, Status>

Returns details about a space. For an example, see Get a space.

Requires authentication. Supports app authentication and user authentication.

pub async fn create_space( &mut self, request: impl IntoRequest<CreateSpaceRequest> ) -> Result<Response<Space>, Status>

Creates a named space. Spaces grouped by topics aren’t supported. For an example, see Create a space.

If you receive the error message ALREADY_EXISTS when creating a space, try a different displayName. An existing space within the Google Workspace organization might already use this display name.

Requires user authentication.

pub async fn set_up_space( &mut self, request: impl IntoRequest<SetUpSpaceRequest> ) -> Result<Response<Space>, Status>

Creates a space and adds specified users to it. The calling user is automatically added to the space, and shouldn’t be specified as a membership in the request. For an example, see Set up a space.

To specify the human members to add, add memberships with the appropriate member.name in the SetUpSpaceRequest. To add a human user, use users/{user}, where {user} can be the email address for the user. For users in the same Workspace organization {user} can also be the id for the person from the People API, or the id for the user in the Directory API. For example, if the People API Person profile ID for user@example.com is 123456789, you can add the user to the space by setting the membership.member.name to users/user@example.com or users/123456789.

For a space or group chat, if the caller blocks or is blocked by some members, then those members aren’t added to the created space.

To create a direct message (DM) between the calling user and another human user, specify exactly one membership to represent the human user. If one user blocks the other, the request fails and the DM isn’t created.

To create a DM between the calling user and the calling app, set Space.singleUserBotDm to true and don’t specify any memberships. You can only use this method to set up a DM with the calling app. To add the calling app as a member of a space or an existing DM between two human users, see create a membership.

If a DM already exists between two users, even when one user blocks the other at the time a request is made, then the existing DM is returned.

Spaces with threaded replies aren’t supported. If you receive the error message ALREADY_EXISTS when setting up a space, try a different displayName. An existing space within the Google Workspace organization might already use this display name.

Requires user authentication.

pub async fn update_space( &mut self, request: impl IntoRequest<UpdateSpaceRequest> ) -> Result<Response<Space>, Status>

Updates a space. For an example, see Update a space.

If you’re updating the displayName field and receive the error message ALREADY_EXISTS, try a different display name.. An existing space within the Google Workspace organization might already use this display name.

Requires user authentication.

pub async fn delete_space( &mut self, request: impl IntoRequest<DeleteSpaceRequest> ) -> Result<Response<()>, Status>

Deletes a named space. Always performs a cascading delete, which means that the space’s child resources—like messages posted in the space and memberships in the space—are also deleted. For an example, see Delete a space. Requires user authentication from a user who has permission to delete the space.

pub async fn complete_import_space( &mut self, request: impl IntoRequest<CompleteImportSpaceRequest> ) -> Result<Response<CompleteImportSpaceResponse>, Status>

Completes the import process for the specified space and makes it visible to users. Requires app authentication and domain-wide delegation. For more information, see Authorize Google Chat apps to import data.

pub async fn find_direct_message( &mut self, request: impl IntoRequest<FindDirectMessageRequest> ) -> Result<Response<Space>, Status>

Returns the existing direct message with the specified user. If no direct message space is found, returns a 404 NOT_FOUND error. For an example, see Find a direct message.

With user authentication, returns the direct message space between the specified user and the authenticated user.

With app authentication, returns the direct message space between the specified user and the calling Chat app.

Requires user authentication or app authentication.

pub async fn create_membership( &mut self, request: impl IntoRequest<CreateMembershipRequest> ) -> Result<Response<Membership>, Status>

Creates a human membership or app membership for the calling app. Creating memberships for other apps isn’t supported. For an example, see Create a membership. When creating a membership, if the specified member has their auto-accept policy turned off, then they’re invited, and must accept the space invitation before joining. Otherwise, creating a membership adds the member directly to the specified space. Requires user authentication.

To specify the member to add, set the membership.member.name in the CreateMembershipRequest:

To add the calling app to a space or a direct message between two human users, use users/app. Unable to add other apps to the space.

To add a human user, use users/{user}, where {user} can be the email address for the user. For users in the same Workspace organization {user} can also be the id for the person from the People API, or the id for the user in the Directory API. For example, if the People API Person profile ID for user@example.com is 123456789, you can add the user to the space by setting the membership.member.name to users/user@example.com or users/123456789.

pub async fn delete_membership( &mut self, request: impl IntoRequest<DeleteMembershipRequest> ) -> Result<Response<Membership>, Status>

Deletes a membership. For an example, see Delete a membership.

Requires user authentication.

pub async fn create_reaction( &mut self, request: impl IntoRequest<CreateReactionRequest> ) -> Result<Response<Reaction>, Status>

Creates a reaction and adds it to a message. For an example, see Create a reaction. Requires user authentication. Only unicode emoji are supported.

pub async fn list_reactions( &mut self, request: impl IntoRequest<ListReactionsRequest> ) -> Result<Response<ListReactionsResponse>, Status>

Lists reactions to a message. For an example, see List reactions. Requires user authentication.

pub async fn delete_reaction( &mut self, request: impl IntoRequest<DeleteReactionRequest> ) -> Result<Response<()>, Status>

Deletes a reaction to a message. For an example, see Delete a reaction. Requires user authentication.