Module google_api_proto::google::chromeos::moblab::v1beta1

Modules

build

Nested message and enum types in Build.

build_service_client

Generated client implementations.

Structs

Build

Resource that represents a build for the given build target, model, milestone and build version. – NEXT_TAG: 9 –

BuildArtifact

Resource that represents a build artifact stored in Google Cloud Storage for the given build target, model, build version and bucket. – NEXT_TAG: 6 –

BuildTarget

Resource that represents a build target. – NEXT_TAG: 2 –

CheckBuildStageStatusRequest

Request message for checking if the build artifact is staged. – NEXT_TAG: 3 –

CheckBuildStageStatusResponse

Response message for checking the stage status of a build artifact. – NEXT_TAG: 4 –

FindMostStableBuildRequest

Request message for finding the most stable build. – NEXT_TAG: 2 –

FindMostStableBuildResponse

Response message for finding the most stable build. – NEXT_TAG: 2 –

ListBuildTargetsRequest

Request message for listing build targets. – NEXT_TAG: 3 –

ListBuildTargetsResponse

Response message for listing build targets. – NEXT_TAG: 4 –

ListBuildsRequest

Request message for listing builds. – NEXT_TAG: 7 –

ListBuildsResponse

Response message for listing builds. – NEXT_TAG: 4 –

ListModelsRequest

Request message for listing models. – NEXT_TAG: 4 –

ListModelsResponse

Response message for listing models. – NEXT_TAG: 4 –

Milestone

Resource that represents a chrome OS milestone. – NEXT_TAG: 2 –

Model

Resource that represents a model. Each model belongs to a build target. For non-unified build, the model name is the same as its build target name. – NEXT_TAG: 2 –

StageBuildMetadata

Metadata message for staging a build artifact. – NEXT_TAG: 4 –

StageBuildRequest

Request message for staging a build artifact. – NEXT_TAG: 3 –

StageBuildResponse

Response message for staging a build artifact. – NEXT_TAG: 2 –