Module google_api_proto::google::cloud::aiplatform::v1::study_spec

Nested message and enum types in StudySpec.

Modules

metric_spec

Nested message and enum types in MetricSpec.

parameter_spec

Nested message and enum types in ParameterSpec.

Structs

ConvexAutomatedStoppingSpec

Configuration for ConvexAutomatedStoppingSpec. When there are enough completed trials (configured by min_measurement_count), for pending trials with enough measurements and steps, the policy first computes an overestimate of the objective value at max_num_steps according to the slope of the incomplete objective value curve. No prediction can be made if the curve is completely flat. If the overestimation is worse than the best objective value of the completed trials, this pending trial will be early-stopped, but a last measurement will be added to the pending trial with max_num_steps and predicted objective value from the autoregression model.

DecayCurveAutomatedStoppingSpec

The decay curve automated stopping rule builds a Gaussian Process Regressor to predict the final objective value of a Trial based on the already completed Trials and the intermediate measurements of the current Trial. Early stopping is requested for the current Trial if there is very low probability to exceed the optimal value found so far.

MedianAutomatedStoppingSpec

The median automated stopping rule stops a pending Trial if the Trial’s best objective_value is strictly below the median ‘performance’ of all completed Trials reported up to the Trial’s last measurement. Currently, ‘performance’ refers to the running average of the objective values reported by the Trial in each measurement.

MetricSpec

Represents a metric to optimize.

ParameterSpec

Represents a single parameter to optimize.

StudyStoppingConfig

The configuration (stopping conditions) for automated stopping of a Study. Conditions include trial budgets, time budgets, and convergence detection.

Enums

Algorithm

The available search algorithms for the Study.

AutomatedStoppingSpec

MeasurementSelectionType

This indicates which measurement to use if/when the service automatically selects the final measurement from previously reported intermediate measurements. Choose this based on two considerations: A) Do you expect your measurements to monotonically improve? If so, choose LAST_MEASUREMENT. On the other hand, if you’re in a situation where your system can “over-train” and you expect the performance to get better for a while but then start declining, choose BEST_MEASUREMENT. B) Are your measurements significantly noisy and/or irreproducible? If so, BEST_MEASUREMENT will tend to be over-optimistic, and it may be better to choose LAST_MEASUREMENT. If both or neither of (A) and (B) apply, it doesn’t matter which selection type is chosen.

ObservationNoise

Describes the noise level of the repeated observations.