Module google_api_proto::google::cloud::apigeeregistry::v1

Modules

instance

Nested message and enum types in Instance.

provisioning_client

Generated client implementations.

registry_client

Generated client implementations.

Structs

Api

A top-level description of an API. Produced by producers and are commitments to provide services.

ApiDeployment

Describes a service running at particular address that provides a particular version of an API. ApiDeployments have revisions which correspond to different configurations of a single deployment in time. Revision identifiers should be updated whenever the served API spec or endpoint address changes.

ApiSpec

Describes a version of an API in a structured way. ApiSpecs provide formal descriptions that consumers can use to use a version. ApiSpec resources are intended to be fully-resolved descriptions of an ApiVersion. When specs consist of multiple files, these should be bundled together (e.g., in a zip archive) and stored as a unit. Multiple specs can exist to provide representations in different API description formats. Synchronization of these representations would be provided by tooling and background services.

ApiVersion

Describes a particular version of an API. ApiVersions are what consumers actually use.

Artifact

Artifacts of resources. Artifacts are unique (single-value) per resource and are used to store metadata that is too large or numerous to be stored directly on the resource. Since artifacts are stored separately from parent resources, they should generally be used for metadata that is needed infrequently, i.e., not for display in primary views of the resource but perhaps displayed or downloaded upon request. The ListArtifacts method allows artifacts to be quickly enumerated and checked for presence without downloading their (potentially-large) contents.

CreateApiDeploymentRequest

Request message for CreateApiDeployment.

CreateApiRequest

Request message for CreateApi.

CreateApiSpecRequest

Request message for CreateApiSpec.

CreateApiVersionRequest

Request message for CreateApiVersion.

CreateArtifactRequest

Request message for CreateArtifact.

CreateInstanceRequest

Request message for CreateInstance.

DeleteApiDeploymentRequest

Request message for DeleteApiDeployment.

DeleteApiDeploymentRevisionRequest

Request message for DeleteApiDeploymentRevision.

DeleteApiRequest

Request message for DeleteApi.

DeleteApiSpecRequest

Request message for DeleteApiSpec.

DeleteApiSpecRevisionRequest

Request message for DeleteApiSpecRevision.

DeleteApiVersionRequest

Request message for DeleteApiVersion.

DeleteArtifactRequest

Request message for DeleteArtifact.

DeleteInstanceRequest

Request message for DeleteInstance.

GetApiDeploymentRequest

Request message for GetApiDeployment.

GetApiRequest

Request message for GetApi.

GetApiSpecContentsRequest

Request message for GetApiSpecContents.

GetApiSpecRequest

Request message for GetApiSpec.

GetApiVersionRequest

Request message for GetApiVersion.

GetArtifactContentsRequest

Request message for GetArtifactContents.

GetArtifactRequest

Request message for GetArtifact.

GetInstanceRequest

Request message for GetInstance.

Instance

An Instance represents the instance resources of the Registry. Currently, only one instance is allowed for each project.

ListApiDeploymentRevisionsRequest

Request message for ListApiDeploymentRevisions.

ListApiDeploymentRevisionsResponse

Response message for ListApiDeploymentRevisionsResponse.

ListApiDeploymentsRequest

Request message for ListApiDeployments.

ListApiDeploymentsResponse

Response message for ListApiDeployments.

ListApiSpecRevisionsRequest

Request message for ListApiSpecRevisions.

ListApiSpecRevisionsResponse

Response message for ListApiSpecRevisionsResponse.

ListApiSpecsRequest

Request message for ListApiSpecs.

ListApiSpecsResponse

Response message for ListApiSpecs.

ListApiVersionsRequest

Request message for ListApiVersions.

ListApiVersionsResponse

Response message for ListApiVersions.

ListApisRequest

Request message for ListApis.

ListApisResponse

Response message for ListApis.

ListArtifactsRequest

Request message for ListArtifacts.

ListArtifactsResponse

Response message for ListArtifacts.

OperationMetadata

Represents the metadata of the long-running operation.

ReplaceArtifactRequest

Request message for ReplaceArtifact.

RollbackApiDeploymentRequest

Request message for RollbackApiDeployment.

RollbackApiSpecRequest

Request message for RollbackApiSpec.

TagApiDeploymentRevisionRequest

Request message for TagApiDeploymentRevision.

TagApiSpecRevisionRequest

Request message for TagApiSpecRevision.

UpdateApiDeploymentRequest

Request message for UpdateApiDeployment.

UpdateApiRequest

Request message for UpdateApi.

UpdateApiSpecRequest

Request message for UpdateApiSpec.

UpdateApiVersionRequest

Request message for UpdateApiVersion.