Module google_api_proto::google::cloud::asset::v1p2beta1

Modules

asset_service_client

Generated client implementations.

feed_output_config

Nested message and enum types in FeedOutputConfig.

gcs_destination

Nested message and enum types in GcsDestination.

output_config

Nested message and enum types in OutputConfig.

Structs

Asset

Cloud asset. This includes all Google Cloud Platform resources, Cloud IAM policies, and other non-GCP assets.

CreateFeedRequest

Create asset feed request.

DeleteFeedRequest

Feed

An asset feed used to export asset updates to a destinations. An asset feed filter controls what updates are exported. The asset feed must be created within a project, organization, or folder. Supported destinations are: Cloud Pub/Sub topics.

FeedOutputConfig

Output configuration for asset feed destination.

GcsDestination

A Cloud Storage location.

GetFeedRequest

Get asset feed request.

ListFeedsRequest

List asset feeds request.

ListFeedsResponse

OutputConfig

Output configuration for export assets destination.

PubsubDestination

A Cloud Pubsub destination.

Resource

Representation of a cloud resource.

TemporalAsset

Temporal asset. In addition to the asset, the temporal asset includes the status of the asset and valid from and to time of it.

TimeWindow

A time window of (start_time, end_time].

UpdateFeedRequest

Update asset feed request.

Enums

ContentType

Asset content type.