Module google_api_proto::google::cloud::audit::big_query_audit_metadata

Nested message and enum types in BigQueryAuditMetadata.

Modules

dataset_change

Nested message and enum types in DatasetChange.

dataset_creation

Nested message and enum types in DatasetCreation.

dataset_deletion

Nested message and enum types in DatasetDeletion.

first_party_app_metadata

Nested message and enum types in FirstPartyAppMetadata.

job_config

Nested message and enum types in JobConfig.

job_deletion

Nested message and enum types in JobDeletion.

job_insertion

Nested message and enum types in JobInsertion.

job_stats

Nested message and enum types in JobStats.

model_creation

Nested message and enum types in ModelCreation.

model_data_change

Nested message and enum types in ModelDataChange.

model_data_read

Nested message and enum types in ModelDataRead.

model_deletion

Nested message and enum types in ModelDeletion.

model_metadata_change

Nested message and enum types in ModelMetadataChange.

routine_change

Nested message and enum types in RoutineChange.

routine_creation

Nested message and enum types in RoutineCreation.

routine_deletion

Nested message and enum types in RoutineDeletion.

table_change

Nested message and enum types in TableChange.

table_creation

Nested message and enum types in TableCreation.

table_data_change

Nested message and enum types in TableDataChange.

table_data_read

Nested message and enum types in TableDataRead.

table_deletion

Nested message and enum types in TableDeletion.

unlink_dataset

Nested message and enum types in UnlinkDataset.

Structs

BigQueryAcl

An access control list.

Dataset

BigQuery dataset.

DatasetChange

Dataset change event.

DatasetCreation

Dataset creation event.

DatasetDeletion

Dataset deletion event.

EncryptionInfo

Encryption properties for a table or a job

EntityInfo

User-provided metadata for an entity, for e.g. dataset, table or model.

FirstPartyAppMetadata

First party (Google) application specific request metadata.

Job

BigQuery job.

JobChange

Job state change event.

JobConfig

Job configuration. See the Jobs API resource for more details on individual fields.

JobDeletion

Job deletion event.

JobInsertion

Job insertion event.

JobStats

Job statistics.

JobStatus

Status of a job.

Model

Trained BigQuery ML model.

ModelCreation

Model creation event.

ModelDataChange

Model data change event.

ModelDataRead

Model data read event.

ModelDeletion

Model deletion event.

ModelMetadataChange

Model metadata change event.

Routine

User Defined Function (UDF) or Stored Procedure.

RoutineChange

Routine change event.

RoutineCreation

Routine creation event.

RoutineDeletion

Routine deletion event.

RowAccessPolicy

BigQuery row access policy.

RowAccessPolicyChange

Row access policy change event.

RowAccessPolicyCreation

Row access policy creation event.

RowAccessPolicyDeletion

Row access policy deletion event.

SheetsMetadata

Google Sheets specific request metadata.

Table

BigQuery table.

TableChange

Table metadata change event.

TableCreation

Table creation event.

TableDataChange

Table data change event.

TableDataRead

Table data read event.

TableDefinition

Definition of an external data source used in a query.

TableDeletion

Table deletion event.

TableViewDefinition

View definition.

UnlinkDataset

Unlink linked dataset from its source dataset event

Enums

CreateDisposition

Describes whether a job should create a destination table if it doesn’t exist.

Event

BigQuery event information.

JobState

State of a job.

OperationType

Table copy job operation type.

QueryStatementType

Type of the statement (e.g. SELECT, INSERT, CREATE_TABLE, CREATE_MODEL..)

WriteDisposition

Describes whether a job should overwrite or append the existing destination table if it already exists.