Module google_api_proto::google::cloud::baremetalsolution::v2

Modules

bare_metal_solution_client

Generated client implementations.

instance

Nested message and enum types in Instance.

logical_interface

Nested message and enum types in LogicalInterface.

lun

Nested message and enum types in Lun.

network

Nested message and enum types in Network.

nfs_share

Nested message and enum types in NfsShare.

server_network_template

Nested message and enum types in ServerNetworkTemplate.

volume

Nested message and enum types in Volume.

vrf

Nested message and enum types in VRF.

Structs

DetachLunRequest

Message for detach specific LUN from an Instance.

GetInstanceRequest

Message for requesting server information.

GetLunRequest

Message for requesting storage lun information.

GetNetworkRequest

Message for requesting network information.

GetNfsShareRequest

Message for requesting NFS share information.

GetVolumeRequest

Message for requesting storage volume information.

Instance

A server.

ListInstancesRequest

Message for requesting the list of servers.

ListInstancesResponse

Response message for the list of servers.

ListLunsRequest

Message for requesting a list of storage volume luns.

ListLunsResponse

Response message containing the list of storage volume luns.

ListNetworkUsageRequest

Request to get networks with IPs.

ListNetworkUsageResponse

Response with Networks with IPs

ListNetworksRequest

Message for requesting a list of networks.

ListNetworksResponse

Response message containing the list of networks.

ListNfsSharesRequest

Message for requesting a list of NFS shares.

ListNfsSharesResponse

Response message containing the list of NFS shares.

ListVolumesRequest

Message for requesting a list of storage volumes.

ListVolumesResponse

Response message containing the list of storage volumes.

LogicalInterface

Each logical interface represents a logical abstraction of the underlying physical interface (for eg. bond, nic) of the instance. Each logical interface can effectively map to multiple network-IP pairs and still be mapped to one underlying physical interface.

Lun

A storage volume logical unit number (LUN).

Network

A Network.

NetworkAddressReservation

A reservation of one or more addresses in a network.

NetworkUsage

Network with all used IP addresses.

NfsShare

An NFS share.

OperationMetadata

Represents the metadata from a long-running operation.

ResetInstanceRequest

Message requesting to reset a server.

ResetInstanceResponse

Response message from resetting a server.

ResizeVolumeRequest

Request for emergency resize Volume.

ServerNetworkTemplate

Network template.

StartInstanceRequest

Message requesting to start a server.

StartInstanceResponse

Response message from starting a server.

StopInstanceRequest

Message requesting to stop a server.

StopInstanceResponse

Response message from stopping a server.

UpdateInstanceRequest

Message requesting to updating a server.

UpdateNetworkRequest

Message requesting to updating a network.

UpdateNfsShareRequest

Message requesting to updating a NFS share.

UpdateVolumeRequest

Message for updating a volume.

Volume

A storage volume.

Vrf

A network VRF.