Module google_api_proto::google::cloud::baremetalsolution::v2

Modules

bare_metal_solution_client

Generated client implementations.

instance

Nested message and enum types in Instance.

lun

Nested message and enum types in Lun.

network

Nested message and enum types in Network.

snapshot_schedule_policy

Nested message and enum types in SnapshotSchedulePolicy.

volume

Nested message and enum types in Volume.

vrf

Nested message and enum types in VRF.

Structs

CreateSnapshotSchedulePolicyRequest

Message for creating a snapshot schedule policy in a project.

CreateVolumeSnapshotRequest

Message for creating a volume snapshot.

DeleteSnapshotSchedulePolicyRequest

Message for deleting a snapshot schedule policy in a project.

DeleteVolumeSnapshotRequest

Message for deleting named Volume snapshot.

GetInstanceRequest

Message for requesting server information.

GetLunRequest

Message for requesting storage lun information.

GetNetworkRequest

Message for requesting network information.

GetSnapshotSchedulePolicyRequest

Message for requesting snapshot schedule policy information.

GetVolumeRequest

Message for requesting storage volume information.

GetVolumeSnapshotRequest

Message for requesting storage volume snapshot information.

Instance

A server.

ListInstancesRequest

Message for requesting the list of servers.

ListInstancesResponse

Response message for the list of servers.

ListLunsRequest

Message for requesting a list of storage volume luns.

ListLunsResponse

Response message containing the list of storage volume luns.

ListNetworksRequest

Message for requesting a list of networks.

ListNetworksResponse

Response message containing the list of networks.

ListSnapshotSchedulePoliciesRequest

Message for requesting a list of snapshot schedule policies.

ListSnapshotSchedulePoliciesResponse

Response message containing the list of snapshot schedule policies.

ListVolumeSnapshotsRequest

Message for requesting a list of storage volume snapshots.

ListVolumeSnapshotsResponse

Response message containing the list of storage volume snapshots.

ListVolumesRequest

Message for requesting a list of storage volumes.

ListVolumesResponse

Response message containing the list of storage volumes.

Lun

A storage volume logical unit number (LUN).

Network

A Network.

OperationMetadata

Represents the metadata from a long-running operation.

ResetInstanceRequest

Message requesting to reset a server.

ResetInstanceResponse

Response message from resetting a server.

RestoreVolumeSnapshotRequest

Message for restoring a volume snapshot.

SnapshotSchedulePolicy

A snapshot schedule policy.

UpdateSnapshotSchedulePolicyRequest

Message for updating a snapshot schedule policy in a project.

UpdateVolumeRequest

Message for updating a volume.

Volume

A storage volume.

VolumeSnapshot

Snapshot registered for a given storage volume.

Vrf

A network VRF.