Module google_api_proto::google::cloud::bigquery::dataexchange::v1beta1

Modules

analytics_hub_service_client

Generated client implementations.

listing

Nested message and enum types in Listing.

subscribe_listing_request

Nested message and enum types in SubscribeListingRequest.

Structs

CreateDataExchangeRequest

Message for creating a data exchange.

CreateListingRequest

Message for creating a listing.

DataExchange

A data exchange is a container that lets you share data. Along with the descriptive information about the data exchange, it contains listings that reference shared datasets.

DataProvider

Contains details of the data provider.

DeleteDataExchangeRequest

Message for deleting a data exchange.

DeleteListingRequest

Message for deleting a listing.

DestinationDataset

Defines the destination bigquery dataset.

DestinationDatasetReference

Contains the reference that identifies a destination bigquery dataset.

GetDataExchangeRequest

Message for getting a data exchange.

GetListingRequest

Message for getting a listing.

ListDataExchangesRequest

Message for requesting the list of data exchanges.

ListDataExchangesResponse

Message for response to the list of data exchanges.

ListListingsRequest

Message for requesting the list of listings.

ListListingsResponse

Message for response to the list of Listings.

ListOrgDataExchangesRequest

Message for requesting the list of data exchanges from projects in an organization and location.

ListOrgDataExchangesResponse

Message for response to listing data exchanges in an organization and location.

Listing

A listing is what gets published into a data exchange that a subscriber can subscribe to. It contains a reference to the data source along with descriptive information that will help subscribers find and subscribe the data.

Publisher

Contains details of the listing publisher.

SubscribeListingRequest

Message for subscribing to a listing.

SubscribeListingResponse

Message for response when you subscribe to a listing.

UpdateDataExchangeRequest

Message for updating a data exchange.

UpdateListingRequest

Message for updating a Listing.