Module google_api_proto::google::cloud::bigquery::dataexchange::v1beta1

Modules

analytics_hub_service_client

Generated client implementations.

listing

Nested message and enum types in Listing.

subscribe_listing_request

Nested message and enum types in SubscribeListingRequest.

Structs

CreateDataExchangeRequest

Message for creating a DataExchange.

CreateListingRequest

Message for creating a Listing.

DataExchange

A data exchange is a container that enables data sharing. It contains a set of listings of the data sources along with descriptive information of the data exchange.

DataProvider

Contains details of the Data Provider.

DeleteDataExchangeRequest

Message for deleting a DataExchange.

DeleteListingRequest

Message for deleting a Listing.

DestinationDataset

Defines the Destination BigQuery Dataset.

DestinationDatasetReference

Defines the Destination BigQuery Dataset Reference.

GetDataExchangeRequest

Message for getting a DataExchange.

GetListingRequest

Message for getting a Listing.

ListDataExchangesRequest

Message for requesting list of DataExchanges.

ListDataExchangesResponse

Message for response to listing DataExchanges.

ListListingsRequest

Message for requesting list of Listings.

ListListingsResponse

Message for response to listing Listings.

ListOrgDataExchangesRequest

Message for requesting list of DataExchanges from projects in an organization and location.

ListOrgDataExchangesResponse

Message for response to listing DataExchanges in an organization and location.

Listing

A listing is what gets published into a data exchange that a subscriber can subscribe to. It contains a reference to the data source along with descriptive information that will help subscribers find and subscribe the data.

Publisher

Contains details of the Publisher.

SubscribeListingRequest

Message for subscribing a Listing.

SubscribeListingResponse

Message for response to subscribing a Listing. Empty for now.

UpdateDataExchangeRequest

Message for updating a DataExchange.

UpdateListingRequest

Message for updating a Listing.