Module google_api_proto::google::cloud::certificatemanager::v1

Modules

certificate

Nested message and enum types in Certificate.

certificate_issuance_config

Nested message and enum types in CertificateIssuanceConfig.

certificate_manager_client

Generated client implementations.

certificate_map

Nested message and enum types in CertificateMap.

certificate_map_entry

Nested message and enum types in CertificateMapEntry.

dns_authorization

Nested message and enum types in DnsAuthorization.

Structs

Certificate

Defines TLS certificate.

CertificateIssuanceConfig

CertificateIssuanceConfig specifies how to issue and manage a certificate.

CertificateMap

Defines a collection of certificate configurations.

CertificateMapEntry

Defines a certificate map entry.

CreateCertificateIssuanceConfigRequest

Request for the CreateCertificateIssuanceConfig method.

CreateCertificateMapEntryRequest

Request for the CreateCertificateMapEntry method.

CreateCertificateMapRequest

Request for the CreateCertificateMap method.

CreateCertificateRequest

Request for the CreateCertificate method.

CreateDnsAuthorizationRequest

Request for the CreateDnsAuthorization method.

DeleteCertificateIssuanceConfigRequest

Request for the DeleteCertificateIssuanceConfig method.

DeleteCertificateMapEntryRequest

Request for the DeleteCertificateMapEntry method.

DeleteCertificateMapRequest

Request for the DeleteCertificateMap method.

DeleteCertificateRequest

Request for the DeleteCertificate method.

DeleteDnsAuthorizationRequest

Request for the DeleteDnsAuthorization method.

DnsAuthorization

A DnsAuthorization resource describes a way to perform domain authorization for certificate issuance.

GetCertificateIssuanceConfigRequest

Request for the GetCertificateIssuanceConfig method.

GetCertificateMapEntryRequest

Request for the GetCertificateMapEntry method.

GetCertificateMapRequest

Request for the GetCertificateMap method.

GetCertificateRequest

Request for the GetCertificate method.

GetDnsAuthorizationRequest

Request for the GetDnsAuthorization method.

ListCertificateIssuanceConfigsRequest

Request for the ListCertificateIssuanceConfigs method.

ListCertificateIssuanceConfigsResponse

Response for the ListCertificateIssuanceConfigs method.

ListCertificateMapEntriesRequest

Request for the ListCertificateMapEntries method.

ListCertificateMapEntriesResponse

Response for the ListCertificateMapEntries method.

ListCertificateMapsRequest

Request for the ListCertificateMaps method.

ListCertificateMapsResponse

Response for the ListCertificateMaps method.

ListCertificatesRequest

Request for the ListCertificates method.

ListCertificatesResponse

Response for the ListCertificates method.

ListDnsAuthorizationsRequest

Request for the ListDnsAuthorizations method.

ListDnsAuthorizationsResponse

Response for the ListDnsAuthorizations method.

OperationMetadata

Represents the metadata of the long-running operation. Output only.

UpdateCertificateMapEntryRequest

Request for the UpdateCertificateMapEntry method.

UpdateCertificateMapRequest

Request for the UpdateCertificateMap method.

UpdateCertificateRequest

Request for the UpdateCertificate method.

UpdateDnsAuthorizationRequest

Request for the UpdateDnsAuthorization method.

Enums

ServingState

Defines set of serving states associated with a resource.