Module google_api_proto::google::cloud::channel::v1

Modules

cloud_channel_reports_service_client

Generated client implementations.

cloud_channel_service_client

Generated client implementations.

cloud_identity_info

Nested message and enum types in CloudIdentityInfo.

column

Nested message and enum types in Column.

customer_event

Nested message and enum types in CustomerEvent.

edu_data

Nested message and enum types in EduData.

entitlement

Nested message and enum types in Entitlement.

entitlement_change

Nested message and enum types in EntitlementChange.

entitlement_event

Nested message and enum types in EntitlementEvent.

import_customer_request

Nested message and enum types in ImportCustomerRequest.

list_purchasable_offers_request

Nested message and enum types in ListPurchasableOffersRequest.

list_purchasable_skus_request

Nested message and enum types in ListPurchasableSkusRequest.

list_transferable_offers_request

Nested message and enum types in ListTransferableOffersRequest.

list_transferable_skus_request

Nested message and enum types in ListTransferableSkusRequest.

operation_metadata

Nested message and enum types in OperationMetadata.

parameter_definition

Nested message and enum types in ParameterDefinition.

report_status

Nested message and enum types in ReportStatus.

report_value

Nested message and enum types in ReportValue.

repricing_adjustment

Nested message and enum types in RepricingAdjustment.

repricing_condition

Nested message and enum types in RepricingCondition.

repricing_config

Nested message and enum types in RepricingConfig.

subscriber_event

Nested message and enum types in SubscriberEvent.

transfer_eligibility

Nested message and enum types in TransferEligibility.

value

Nested message and enum types in Value.

Structs

ActivateEntitlementRequest

Request message for [CloudChannelService.ActivateEntitlement][google.cloud.channel.v1.CloudChannelService.ActivateEntitlement].

AdminUser

Information needed to create an Admin User for Google Workspace.

AssociationInfo

Association links that an entitlement has to other entitlements.

BillableSku

Represents the Billable SKU information.

BillingAccount

Represents a billing account.

BillingAccountPurchaseInfo

Represents a billing account that can be used to make a purchase.

CancelEntitlementRequest

Request message for [CloudChannelService.CancelEntitlement][google.cloud.channel.v1.CloudChannelService.CancelEntitlement].

ChangeOfferRequest

Request message for [CloudChannelService.ChangeOffer][google.cloud.channel.v1.CloudChannelService.ChangeOffer].

ChangeParametersRequest

Request message for [CloudChannelService.ChangeParametersRequest][].

ChangeRenewalSettingsRequest

Request message for [CloudChannelService.ChangeRenewalSettings][google.cloud.channel.v1.CloudChannelService.ChangeRenewalSettings].

ChannelPartnerLink

Entity representing a link between distributors and their indirect resellers in an n-tier resale channel.

ChannelPartnerRepricingConfig

Configuration for how a distributor will rebill a channel partner (also known as a distributor-authorized reseller).

CheckCloudIdentityAccountsExistRequest

Request message for [CloudChannelService.CheckCloudIdentityAccountsExist][google.cloud.channel.v1.CloudChannelService.CheckCloudIdentityAccountsExist].

CheckCloudIdentityAccountsExistResponse

Response message for [CloudChannelService.CheckCloudIdentityAccountsExist][google.cloud.channel.v1.CloudChannelService.CheckCloudIdentityAccountsExist].

CloudIdentityCustomerAccount

Entity representing a Cloud Identity account that may be associated with a Channel Services API partner.

CloudIdentityInfo

Cloud Identity information for the Cloud Channel Customer.

Column

The definition of a report column. Specifies the data properties in the corresponding position of the report rows.

CommitmentSettings

Commitment settings for commitment-based offers.

ConditionalOverride

Specifies the override to conditionally apply.

Constraints

Represents the constraints for buying the Offer.

ContactInfo

Contact information for a customer account.

CreateChannelPartnerLinkRequest

Request message for [CloudChannelService.CreateChannelPartnerLink][google.cloud.channel.v1.CloudChannelService.CreateChannelPartnerLink]

CreateChannelPartnerRepricingConfigRequest

Request message for [CloudChannelService.CreateChannelPartnerRepricingConfig][google.cloud.channel.v1.CloudChannelService.CreateChannelPartnerRepricingConfig].

CreateCustomerRepricingConfigRequest

Request message for [CloudChannelService.CreateCustomerRepricingConfig][google.cloud.channel.v1.CloudChannelService.CreateCustomerRepricingConfig].

CreateCustomerRequest

Request message for [CloudChannelService.CreateCustomer][google.cloud.channel.v1.CloudChannelService.CreateCustomer]

CreateEntitlementRequest

Request message for [CloudChannelService.CreateEntitlement][google.cloud.channel.v1.CloudChannelService.CreateEntitlement]

Customer

Entity representing a customer of a reseller or distributor.

CustomerConstraints

Represents constraints required to purchase the Offer for a customer.

CustomerEvent

Represents Pub/Sub message content describing customer update.

CustomerRepricingConfig

Configuration for how a reseller will reprice a Customer.

DateRange

A representation of usage or invoice date ranges.

DeleteChannelPartnerRepricingConfigRequest

Request message for DeleteChannelPartnerRepricingConfig.

DeleteCustomerRepricingConfigRequest

Request message for [CloudChannelService.DeleteCustomerRepricingConfig][google.cloud.channel.v1.CloudChannelService.DeleteCustomerRepricingConfig].

DeleteCustomerRequest

Request message for [CloudChannelService.DeleteCustomer][google.cloud.channel.v1.CloudChannelService.DeleteCustomer].

EduData

Required Edu Attributes

Entitlement

An entitlement is a representation of a customer’s ability to use a service.

EntitlementChange

Change event entry for Entitlement order history

EntitlementEvent

Represents Pub/Sub message content describing entitlement update.

FetchReportResultsRequest

Request message for [CloudChannelReportsService.FetchReportResults][google.cloud.channel.v1.CloudChannelReportsService.FetchReportResults].

FetchReportResultsResponse

Response message for [CloudChannelReportsService.FetchReportResults][google.cloud.channel.v1.CloudChannelReportsService.FetchReportResults]. Contains a tabular representation of the report results.

GetChannelPartnerLinkRequest

Request message for [CloudChannelService.GetChannelPartnerLink][google.cloud.channel.v1.CloudChannelService.GetChannelPartnerLink].

GetChannelPartnerRepricingConfigRequest

Request message for [CloudChannelService.GetChannelPartnerRepricingConfig][google.cloud.channel.v1.CloudChannelService.GetChannelPartnerRepricingConfig]

GetCustomerRepricingConfigRequest

Request message for [CloudChannelService.GetCustomerRepricingConfig][google.cloud.channel.v1.CloudChannelService.GetCustomerRepricingConfig].

GetCustomerRequest

Request message for [CloudChannelService.GetCustomer][google.cloud.channel.v1.CloudChannelService.GetCustomer].

GetEntitlementRequest

Request message for [CloudChannelService.GetEntitlement][google.cloud.channel.v1.CloudChannelService.GetEntitlement].

ImportCustomerRequest

Request message for [CloudChannelService.ImportCustomer][google.cloud.channel.v1.CloudChannelService.ImportCustomer]

ListChannelPartnerLinksRequest

Request message for [CloudChannelService.ListChannelPartnerLinks][google.cloud.channel.v1.CloudChannelService.ListChannelPartnerLinks]

ListChannelPartnerLinksResponse

Response message for [CloudChannelService.ListChannelPartnerLinks][google.cloud.channel.v1.CloudChannelService.ListChannelPartnerLinks].

ListChannelPartnerRepricingConfigsRequest

Request message for [CloudChannelService.ListChannelPartnerRepricingConfigs][google.cloud.channel.v1.CloudChannelService.ListChannelPartnerRepricingConfigs].

ListChannelPartnerRepricingConfigsResponse

Response message for [CloudChannelService.ListChannelPartnerRepricingConfigs][google.cloud.channel.v1.CloudChannelService.ListChannelPartnerRepricingConfigs].

ListCustomerRepricingConfigsRequest

Request message for [CloudChannelService.ListCustomerRepricingConfigs][google.cloud.channel.v1.CloudChannelService.ListCustomerRepricingConfigs].

ListCustomerRepricingConfigsResponse

Response message for [CloudChannelService.ListCustomerRepricingConfigs][google.cloud.channel.v1.CloudChannelService.ListCustomerRepricingConfigs].

ListCustomersRequest

Request message for [CloudChannelService.ListCustomers][google.cloud.channel.v1.CloudChannelService.ListCustomers]

ListCustomersResponse

Response message for [CloudChannelService.ListCustomers][google.cloud.channel.v1.CloudChannelService.ListCustomers].

ListEntitlementChangesRequest

Request message for [CloudChannelService.ListEntitlementChanges][google.cloud.channel.v1.CloudChannelService.ListEntitlementChanges]

ListEntitlementChangesResponse

Response message for [CloudChannelService.ListEntitlementChanges][google.cloud.channel.v1.CloudChannelService.ListEntitlementChanges]

ListEntitlementsRequest

Request message for [CloudChannelService.ListEntitlements][google.cloud.channel.v1.CloudChannelService.ListEntitlements]

ListEntitlementsResponse

Response message for [CloudChannelService.ListEntitlements][google.cloud.channel.v1.CloudChannelService.ListEntitlements].

ListOffersRequest

Request message for ListOffers.

ListOffersResponse

Response message for ListOffers.

ListProductsRequest

Request message for ListProducts.

ListProductsResponse

Response message for ListProducts.

ListPurchasableOffersRequest

Request message for ListPurchasableOffers.

ListPurchasableOffersResponse

Response message for ListPurchasableOffers.

ListPurchasableSkusRequest

Request message for ListPurchasableSkus.

ListPurchasableSkusResponse

Response message for ListPurchasableSkus.

ListReportsRequest

Request message for [CloudChannelReportsService.ListReports][google.cloud.channel.v1.CloudChannelReportsService.ListReports].

ListReportsResponse

Response message for [CloudChannelReportsService.ListReports][google.cloud.channel.v1.CloudChannelReportsService.ListReports].

ListSkuGroupBillableSkusRequest

Request message for ListSkuGroupBillableSkus.

ListSkuGroupBillableSkusResponse

Response message for ListSkuGroupBillableSkus.

ListSkuGroupsRequest

Request message for ListSkuGroups.

ListSkuGroupsResponse

Response message for ListSkuGroups.

ListSkusRequest

Request message for ListSkus.

ListSkusResponse

Response message for ListSkus.

ListSubscribersRequest

Request Message for ListSubscribers.

ListSubscribersResponse

Response Message for ListSubscribers.

ListTransferableOffersRequest

Request message for [CloudChannelService.ListTransferableOffers][google.cloud.channel.v1.CloudChannelService.ListTransferableOffers]

ListTransferableOffersResponse

Response message for [CloudChannelService.ListTransferableOffers][google.cloud.channel.v1.CloudChannelService.ListTransferableOffers].

ListTransferableSkusRequest

Request message for [CloudChannelService.ListTransferableSkus][google.cloud.channel.v1.CloudChannelService.ListTransferableSkus]

ListTransferableSkusResponse

Response message for [CloudChannelService.ListTransferableSkus][google.cloud.channel.v1.CloudChannelService.ListTransferableSkus].

LookupOfferRequest

Request message for LookupOffer.

MarketingInfo

Represents the marketing information for a Product, SKU or Offer.

Media

Represents media information.

Offer

Represents an offer made to resellers for purchase. An offer is associated with a [Sku][google.cloud.channel.v1.Sku], has a plan for payment, a price, and defines the constraints for buying.

OperationMetadata

Provides contextual information about a [google.longrunning.Operation][google.longrunning.Operation].

Parameter

Definition for extended entitlement parameters.

ParameterDefinition

Parameter’s definition. Specifies what parameter is required to use the current Offer to purchase.

PercentageAdjustment

An adjustment that applies a flat markup or markdown to an entire bill.

Period

Represents period in days/months/years.

Plan

The payment plan for the Offer. Describes how to make a payment.

Price

Represents the price of the Offer.

PriceByResource

Represents price by resource type.

PricePhase

Specifies the price by the duration of months. For example, a 20% discount for the first six months, then a 10% discount starting on the seventh month.

PriceTier

Defines price at resource tier level. For example, an offer with following definition :

Product

A Product is the entity a customer uses when placing an order. For example, Google Workspace, Google Voice, etc.

ProvisionCloudIdentityRequest

Request message for [CloudChannelService.ProvisionCloudIdentity][google.cloud.channel.v1.CloudChannelService.ProvisionCloudIdentity]

ProvisionedService

Service provisioned for an entitlement.

PurchasableOffer

Offer that you can purchase for a customer. This is used in the ListPurchasableOffer API response.

PurchasableSku

SKU that you can purchase. This is used in ListPurchasableSku API response.

QueryEligibleBillingAccountsRequest

Request message for QueryEligibleBillingAccounts.

QueryEligibleBillingAccountsResponse

Response message for QueryEligibleBillingAccounts.

RegisterSubscriberRequest

Request Message for RegisterSubscriber.

RegisterSubscriberResponse

Response Message for RegisterSubscriber.

RenewalSettings

Renewal settings for renewable Offers.

Report

The ID and description of a report that was used to generate report data. For example, “Google Cloud Daily Spend”, “Google Workspace License Activity”, etc.

ReportJob

The result of a [RunReportJob][] operation. Contains the name to use in [FetchReportResultsRequest.report_job][google.cloud.channel.v1.FetchReportResultsRequest.report_job] and the status of the operation.

ReportResultsMetadata

The features describing the data. Returned by [CloudChannelReportsService.RunReportJob][google.cloud.channel.v1.CloudChannelReportsService.RunReportJob] and [CloudChannelReportsService.FetchReportResults][google.cloud.channel.v1.CloudChannelReportsService.FetchReportResults].

ReportStatus

Status of a report generation process.

ReportValue

A single report value.

RepricingAdjustment

A type that represents the various adjustments you can apply to a bill.

RepricingCondition

Represents the various repricing conditions you can use for a conditional override.

RepricingConfig

Configuration for repricing a Google bill over a period of time.

Row

A row of report values.

RunReportJobRequest

Request message for [CloudChannelReportsService.RunReportJob][google.cloud.channel.v1.CloudChannelReportsService.RunReportJob].

RunReportJobResponse

Response message for [CloudChannelReportsService.RunReportJob][google.cloud.channel.v1.CloudChannelReportsService.RunReportJob].

Sku

Represents a product’s purchasable Stock Keeping Unit (SKU). SKUs represent the different variations of the product. For example, Google Workspace Business Standard and Google Workspace Business Plus are Google Workspace product SKUs.

SkuGroup

Represents the SKU group information.

SkuGroupCondition

A condition that applies the override if a line item SKU is found in the SKU group.

SkuPurchaseGroup

Represents a set of SKUs that must be purchased using the same billing account.

StartPaidServiceRequest

Request message for [CloudChannelService.StartPaidService][google.cloud.channel.v1.CloudChannelService.StartPaidService].

SubscriberEvent

Represents information which resellers will get as part of notification from Pub/Sub.

SuspendEntitlementRequest

Request message for [CloudChannelService.SuspendEntitlement][google.cloud.channel.v1.CloudChannelService.SuspendEntitlement].

TransferEligibility

Specifies transfer eligibility of a SKU.

TransferEntitlementsRequest

Request message for [CloudChannelService.TransferEntitlements][google.cloud.channel.v1.CloudChannelService.TransferEntitlements].

TransferEntitlementsResponse

Response message for [CloudChannelService.TransferEntitlements][google.cloud.channel.v1.CloudChannelService.TransferEntitlements]. This is put in the response field of google.longrunning.Operation.

TransferEntitlementsToGoogleRequest

Request message for [CloudChannelService.TransferEntitlementsToGoogle][google.cloud.channel.v1.CloudChannelService.TransferEntitlementsToGoogle].

TransferableOffer

TransferableOffer represents an Offer that can be used in Transfer. Read-only.

TransferableSku

TransferableSku represents information a reseller needs to view existing provisioned services for a customer that they do not own. Read-only.

TrialSettings

Settings for trial offers.

UnregisterSubscriberRequest

Request Message for UnregisterSubscriber.

UnregisterSubscriberResponse

Response Message for UnregisterSubscriber.

UpdateChannelPartnerLinkRequest

Request message for [CloudChannelService.UpdateChannelPartnerLink][google.cloud.channel.v1.CloudChannelService.UpdateChannelPartnerLink]

UpdateChannelPartnerRepricingConfigRequest

Request message for [CloudChannelService.UpdateChannelPartnerRepricingConfig][google.cloud.channel.v1.CloudChannelService.UpdateChannelPartnerRepricingConfig].

UpdateCustomerRepricingConfigRequest

Request message for [CloudChannelService.UpdateCustomerRepricingConfig][google.cloud.channel.v1.CloudChannelService.UpdateCustomerRepricingConfig].

UpdateCustomerRequest

Request message for [CloudChannelService.UpdateCustomer][google.cloud.channel.v1.CloudChannelService.UpdateCustomer].

Value

Data type and value of a parameter.

Enums

ChannelPartnerLinkState

ChannelPartnerLinkState represents state of a channel partner link.

ChannelPartnerLinkView

The level of granularity the [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] will display.

MediaType

Type of media used.

PaymentPlan

Describes how the reseller will be billed.

PaymentType

Specifies when the payment needs to happen.

PeriodType

Period Type.

PromotionalOrderType

Constraints type for Promotional offers.

RebillingBasis

Specifies the different costs that the modified bill can be based on.

ResourceType

Represents the type for a monetizable resource(any entity on which billing happens). For example, this could be MINUTES for Google Voice and GB for Google Drive. One SKU can map to multiple monetizable resources.