Module google_api_proto::google::cloud::clouddms::logging::v1

Modules

connection_profile_event_log

Nested message and enum types in ConnectionProfileEventLog.

logged_connection_profile

Nested message and enum types in LoggedConnectionProfile.

logged_migration_job

Nested message and enum types in LoggedMigrationJob.

logged_private_connection

Nested message and enum types in LoggedPrivateConnection.

migration_job_event_log

Nested message and enum types in MigrationJobEventLog.

my_sql_connection_profile

Nested message and enum types in MySqlConnectionProfile.

oracle_connection_profile

Nested message and enum types in OracleConnectionProfile.

postgre_sql_connection_profile

Nested message and enum types in PostgreSqlConnectionProfile.

private_connection_event_log

Nested message and enum types in PrivateConnectionEventLog.

Structs

CloudSqlConnectionProfile

A CloudSQL connection profile.

ConnectionProfileEventLog

Log definition for Connection Profile event

DatabaseType

A message defining the database engine and provider.

LoggedConnectionProfile

An producer connection profile definition. NEXT_TAG = 8.

LoggedMigrationJob

Migration job as stored in Cloud Logging logs. NEXT_TAG = 36.

LoggedPrivateConnection

The PrivateConnection resource is used to establish private connectivity with the customer’s network.

MigrationJobEventLog

Log definition for Migration Job event

MySqlConnectionProfile

An MySQL database connection profile.

OracleConnectionProfile

An Oracle connection profile.

PostgreSqlConnectionProfile

An PostgreSQL connection profile.

PrivateConnectionEventLog

Log definition for a Private Connection event.

VpcPeeringConfig

The VPC Peering configuration is used to create VPC peering between the data plane and the consumer’s VPC.

Enums

DatabaseEngine

The database engines.

DatabaseProvider

The database providers.