Module google_api_proto::google::cloud::clouddms::v1

Modules

cloud_sql_settings

Nested message and enum types in CloudSqlSettings.

connection_profile

Nested message and enum types in ConnectionProfile.

data_migration_service_client

Generated client implementations.

generate_ssh_script_request

Nested message and enum types in GenerateSshScriptRequest.

migration_job

Nested message and enum types in MigrationJob.

migration_job_verification_error

Nested message and enum types in MigrationJobVerificationError.

sql_acl_entry

Nested message and enum types in SqlAclEntry.

ssl_config

Nested message and enum types in SslConfig.

Structs

CloudSqlConnectionProfile

Specifies required connection parameters, and, optionally, the parameters required to create a Cloud SQL destination database instance.

CloudSqlSettings

Settings for creating a Cloud SQL database instance.

ConnectionProfile

A connection profile definition.

CreateConnectionProfileRequest

Request message for ‘CreateConnectionProfile’ request.

CreateMigrationJobRequest

Request message to create a new Database Migration Service migration job in the specified project and region.

DatabaseType

A message defining the database engine and provider.

DeleteConnectionProfileRequest

Request message for ‘DeleteConnectionProfile’ request.

DeleteMigrationJobRequest

Request message for ‘DeleteMigrationJob’ request.

GenerateSshScriptRequest

Request message for ‘GenerateSshScript’ request.

GetConnectionProfileRequest

Request message for ‘GetConnectionProfile’ request.

GetMigrationJobRequest

Request message for ‘GetMigrationJob’ request.

ListConnectionProfilesRequest

Request message for ‘ListConnectionProfiles’ request.

ListConnectionProfilesResponse

Response message for ‘ListConnectionProfiles’ request.

ListMigrationJobsRequest

Retrieve a list of all migration jobs in a given project and location.

ListMigrationJobsResponse

Response message for ‘ListMigrationJobs’ request.

MigrationJob

Represents a Database Migration Service migration job object.

MigrationJobVerificationError

Error message of a verification Migration job.

MySqlConnectionProfile

Specifies connection parameters required specifically for MySQL databases.

OperationMetadata

Represents the metadata of the long-running operation.

PostgreSqlConnectionProfile

Specifies connection parameters required specifically for PostgreSQL databases.

PromoteMigrationJobRequest

Request message for ‘PromoteMigrationJob’ request.

RestartMigrationJobRequest

Request message for ‘RestartMigrationJob’ request.

ResumeMigrationJobRequest

Request message for ‘ResumeMigrationJob’ request.

ReverseSshConnectivity

The details needed to configure a reverse SSH tunnel between the source and destination databases. These details will be used when calling the generateSshScript method (see https://cloud.google.com/database-migration/docs/reference/rest/v1/projects.locations.migrationJobs/generateSshScript) to produce the script that will help set up the reverse SSH tunnel, and to set up the VPC peering between the Cloud SQL private network and the VPC.

SqlAclEntry

An entry for an Access Control list.

SqlIpConfig

IP Management configuration.

SshScript

Response message for ‘GenerateSshScript’ request.

SslConfig

SSL configuration information.

StartMigrationJobRequest

Request message for ‘StartMigrationJob’ request.

StaticIpConnectivity

The source database will allow incoming connections from the destination database’s public IP. You can retrieve the Cloud SQL instance’s public IP from the Cloud SQL console or using Cloud SQL APIs. No additional configuration is required.

StopMigrationJobRequest

Request message for ‘StopMigrationJob’ request.

UpdateConnectionProfileRequest

Request message for ‘UpdateConnectionProfile’ request.

UpdateMigrationJobRequest

Request message for ‘UpdateMigrationJob’ request.

VerifyMigrationJobRequest

Request message for ‘VerifyMigrationJob’ request.

VmCreationConfig

VM creation configuration message

VmSelectionConfig

VM selection configuration message

VpcPeeringConnectivity

The details of the VPC where the source database is located in Google Cloud. We will use this information to set up the VPC peering connection between Cloud SQL and this VPC.

Enums

DatabaseEngine

The database engine types.

DatabaseProvider

The database providers.