Struct google_api_proto::google::cloud::compute::v1::instances_client::InstancesClient

pub struct InstancesClient<T> { /* private fields */ }

The Instances API.

Implementations

impl<T> InstancesClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> InstancesClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn add_access_config(
&mut self,
request: impl IntoRequest<AddAccessConfigInstanceRequest>
) -> Result<Response<Operation>, Status>

Adds an access config to an instance’s network interface.

pub async fn add_resource_policies(
&mut self,
request: impl IntoRequest<AddResourcePoliciesInstanceRequest>
) -> Result<Response<Operation>, Status>

Adds existing resource policies to an instance. You can only add one policy right now which will be applied to this instance for scheduling live migrations.

pub async fn aggregated_list(
&mut self,
request: impl IntoRequest<AggregatedListInstancesRequest>
) -> Result<Response<InstanceAggregatedList>, Status>

Retrieves an aggregated list of all of the instances in your project across all regions and zones. The performance of this method degrades when a filter is specified on a project that has a very large number of instances.

pub async fn attach_disk(
&mut self,
request: impl IntoRequest<AttachDiskInstanceRequest>
) -> Result<Response<Operation>, Status>

Attaches an existing Disk resource to an instance. You must first create the disk before you can attach it. It is not possible to create and attach a disk at the same time. For more information, read Adding a persistent disk to your instance.

pub async fn bulk_insert(
&mut self,
request: impl IntoRequest<BulkInsertInstanceRequest>
) -> Result<Response<Operation>, Status>

Creates multiple instances. Count specifies the number of instances to create.

pub async fn delete(
&mut self,
request: impl IntoRequest<DeleteInstanceRequest>
) -> Result<Response<Operation>, Status>

Deletes the specified Instance resource. For more information, see Deleting an instance.

pub async fn delete_access_config(
&mut self,
request: impl IntoRequest<DeleteAccessConfigInstanceRequest>
) -> Result<Response<Operation>, Status>

Deletes an access config from an instance’s network interface.

pub async fn detach_disk(
&mut self,
request: impl IntoRequest<DetachDiskInstanceRequest>
) -> Result<Response<Operation>, Status>

Detaches a disk from an instance.

pub async fn get(
&mut self,
request: impl IntoRequest<GetInstanceRequest>
) -> Result<Response<Instance>, Status>

Returns the specified Instance resource. Gets a list of available instances by making a list() request.

pub async fn get_effective_firewalls(
&mut self,
request: impl IntoRequest<GetEffectiveFirewallsInstanceRequest>
) -> Result<Response<InstancesGetEffectiveFirewallsResponse>, Status>

Returns effective firewalls applied to an interface of the instance.

pub async fn get_guest_attributes(
&mut self,
request: impl IntoRequest<GetGuestAttributesInstanceRequest>
) -> Result<Response<GuestAttributes>, Status>

Returns the specified guest attributes entry.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyInstanceRequest>
) -> Result<Response<Policy>, Status>

Gets the access control policy for a resource. May be empty if no such policy or resource exists.

pub async fn get_screenshot(
&mut self,
request: impl IntoRequest<GetScreenshotInstanceRequest>
) -> Result<Response<Screenshot>, Status>

Returns the screenshot from the specified instance.

pub async fn get_serial_port_output(
&mut self,
request: impl IntoRequest<GetSerialPortOutputInstanceRequest>
) -> Result<Response<SerialPortOutput>, Status>

Returns the last 1 MB of serial port output from the specified instance.

pub async fn get_shielded_instance_identity(
&mut self,
request: impl IntoRequest<GetShieldedInstanceIdentityInstanceRequest>
) -> Result<Response<ShieldedInstanceIdentity>, Status>

Returns the Shielded Instance Identity of an instance

pub async fn insert(
&mut self,
request: impl IntoRequest<InsertInstanceRequest>
) -> Result<Response<Operation>, Status>

Creates an instance resource in the specified project using the data included in the request.

pub async fn list(
&mut self,
request: impl IntoRequest<ListInstancesRequest>
) -> Result<Response<InstanceList>, Status>

Retrieves the list of instances contained within the specified zone.

pub async fn list_referrers(
&mut self,
request: impl IntoRequest<ListReferrersInstancesRequest>
) -> Result<Response<InstanceListReferrers>, Status>

Retrieves a list of resources that refer to the VM instance specified in the request. For example, if the VM instance is part of a managed or unmanaged instance group, the referrers list includes the instance group. For more information, read Viewing referrers to VM instances.

pub async fn remove_resource_policies(
&mut self,
request: impl IntoRequest<RemoveResourcePoliciesInstanceRequest>
) -> Result<Response<Operation>, Status>

Removes resource policies from an instance.

pub async fn reset(
&mut self,
request: impl IntoRequest<ResetInstanceRequest>
) -> Result<Response<Operation>, Status>

Performs a reset on the instance. This is a hard reset. The VM does not do a graceful shutdown. For more information, see Resetting an instance.

pub async fn resume(
&mut self,
request: impl IntoRequest<ResumeInstanceRequest>
) -> Result<Response<Operation>, Status>

Resumes an instance that was suspended using the instances().suspend method.

pub async fn send_diagnostic_interrupt(
&mut self,
request: impl IntoRequest<SendDiagnosticInterruptInstanceRequest>
) -> Result<Response<SendDiagnosticInterruptInstanceResponse>, Status>

Sends diagnostic interrupt to the instance.

pub async fn set_deletion_protection(
&mut self,
request: impl IntoRequest<SetDeletionProtectionInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets deletion protection on the instance.

pub async fn set_disk_auto_delete(
&mut self,
request: impl IntoRequest<SetDiskAutoDeleteInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets the auto-delete flag for a disk attached to an instance.

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyInstanceRequest>
) -> Result<Response<Policy>, Status>

Sets the access control policy on the specified resource. Replaces any existing policy.

pub async fn set_labels(
&mut self,
request: impl IntoRequest<SetLabelsInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets labels on an instance. To learn more about labels, read the Labeling Resources documentation.

pub async fn set_machine_resources(
&mut self,
request: impl IntoRequest<SetMachineResourcesInstanceRequest>
) -> Result<Response<Operation>, Status>

Changes the number and/or type of accelerator for a stopped instance to the values specified in the request.

pub async fn set_machine_type(
&mut self,
request: impl IntoRequest<SetMachineTypeInstanceRequest>
) -> Result<Response<Operation>, Status>

Changes the machine type for a stopped instance to the machine type specified in the request.

pub async fn set_metadata(
&mut self,
request: impl IntoRequest<SetMetadataInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets metadata for the specified instance to the data included in the request.

pub async fn set_min_cpu_platform(
&mut self,
request: impl IntoRequest<SetMinCpuPlatformInstanceRequest>
) -> Result<Response<Operation>, Status>

Changes the minimum CPU platform that this instance should use. This method can only be called on a stopped instance. For more information, read Specifying a Minimum CPU Platform.

pub async fn set_scheduling(
&mut self,
request: impl IntoRequest<SetSchedulingInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets an instance’s scheduling options. You can only call this method on a stopped instance, that is, a VM instance that is in a TERMINATED state. See Instance Life Cycle for more information on the possible instance states. For more information about setting scheduling options for a VM, see Set VM availability policies.

pub async fn set_service_account(
&mut self,
request: impl IntoRequest<SetServiceAccountInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets the service account on the instance. For more information, read Changing the service account and access scopes for an instance.

pub async fn set_shielded_instance_integrity_policy(
&mut self,
request: impl IntoRequest<SetShieldedInstanceIntegrityPolicyInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets the Shielded Instance integrity policy for an instance. You can only use this method on a running instance. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.

pub async fn set_tags(
&mut self,
request: impl IntoRequest<SetTagsInstanceRequest>
) -> Result<Response<Operation>, Status>

Sets network tags for the specified instance to the data included in the request.

pub async fn simulate_maintenance_event(
&mut self,
request: impl IntoRequest<SimulateMaintenanceEventInstanceRequest>
) -> Result<Response<Operation>, Status>

Simulates a host maintenance event on a VM. For more information, see Simulate a host maintenance event.

pub async fn start(
&mut self,
request: impl IntoRequest<StartInstanceRequest>
) -> Result<Response<Operation>, Status>

Starts an instance that was stopped using the instances().stop method. For more information, see Restart an instance.

pub async fn start_with_encryption_key(
&mut self,
request: impl IntoRequest<StartWithEncryptionKeyInstanceRequest>
) -> Result<Response<Operation>, Status>

Starts an instance that was stopped using the instances().stop method. For more information, see Restart an instance.

pub async fn stop(
&mut self,
request: impl IntoRequest<StopInstanceRequest>
) -> Result<Response<Operation>, Status>

Stops a running instance, shutting it down cleanly, and allows you to restart the instance at a later time. Stopped instances do not incur VM usage charges while they are stopped. However, resources that the VM is using, such as persistent disks and static IP addresses, will continue to be charged until they are deleted. For more information, see Stopping an instance.

pub async fn suspend(
&mut self,
request: impl IntoRequest<SuspendInstanceRequest>
) -> Result<Response<Operation>, Status>

This method suspends a running instance, saving its state to persistent storage, and allows you to resume the instance at a later time. Suspended instances have no compute costs (cores or RAM), and incur only storage charges for the saved VM memory and localSSD data. Any charged resources the virtual machine was using, such as persistent disks and static IP addresses, will continue to be charged while the instance is suspended. For more information, see Suspending and resuming an instance.

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsInstanceRequest>
) -> Result<Response<TestPermissionsResponse>, Status>

Returns permissions that a caller has on the specified resource.

pub async fn update(
&mut self,
request: impl IntoRequest<UpdateInstanceRequest>
) -> Result<Response<Operation>, Status>

Updates an instance only if the necessary resources are available. This method can update only a specific set of instance properties. See Updating a running instance for a list of updatable instance properties.

pub async fn update_access_config(
&mut self,
request: impl IntoRequest<UpdateAccessConfigInstanceRequest>
) -> Result<Response<Operation>, Status>

Updates the specified access config from an instance’s network interface with the data included in the request. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.

pub async fn update_display_device(
&mut self,
request: impl IntoRequest<UpdateDisplayDeviceInstanceRequest>
) -> Result<Response<Operation>, Status>

Updates the Display config for a VM instance. You can only use this method on a stopped VM instance. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.

pub async fn update_network_interface(
&mut self,
request: impl IntoRequest<UpdateNetworkInterfaceInstanceRequest>
) -> Result<Response<Operation>, Status>

Updates an instance’s network interface. This method can only update an interface’s alias IP range and attached network. See Modifying alias IP ranges for an existing instance for instructions on changing alias IP ranges. See Migrating a VM between networks for instructions on migrating an interface. This method follows PATCH semantics.

pub async fn update_shielded_instance_config(
&mut self,
request: impl IntoRequest<UpdateShieldedInstanceConfigInstanceRequest>
) -> Result<Response<Operation>, Status>

Updates the Shielded Instance config for an instance. You can only use this method on a stopped instance. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.