Enum google_api_proto::google::cloud::compute::v1::quota::Metric

#[repr(i32)]
pub enum Metric {

    UndefinedMetric,
    A2Cpus,
    AffinityGroups,
    Autoscalers,
    BackendBuckets,
    BackendServices,
    C2dCpus,
    C2Cpus,
    C3Cpus,
    Commitments,
    CommittedA2Cpus,
    CommittedC2dCpus,
    CommittedC2Cpus,
    CommittedC3Cpus,
    CommittedCpus,
    CommittedE2Cpus,
    CommittedLicenses,
    CommittedLocalSsdTotalGb,
    CommittedM3Cpus,
    CommittedMemoryOptimizedCpus,
    CommittedN2aCpus,
    CommittedN2dCpus,
    CommittedN2Cpus,
    CommittedNvidiaA10080gbGpus,
    CommittedNvidiaA100Gpus,
    CommittedNvidiaK80Gpus,
    CommittedNvidiaL4Gpus,
    CommittedNvidiaP100Gpus,
    CommittedNvidiaP4Gpus,
    CommittedNvidiaT4Gpus,
    CommittedNvidiaV100Gpus,
    CommittedT2aCpus,
    CommittedT2dCpus,
    Cpus,
    CpusAllRegions,
    DisksTotalGb,
    E2Cpus,
    ExternalManagedForwardingRules,
    ExternalNetworkLbForwardingRules,
    ExternalProtocolForwardingRules,
    ExternalVpnGateways,
    Firewalls,
    ForwardingRules,
    GlobalExternalManagedBackendServices,
    GlobalExternalManagedForwardingRules,
    GlobalExternalProxyLbBackendServices,
    GlobalInternalAddresses,
    GlobalInternalManagedBackendServices,
    GlobalInternalTrafficDirectorBackendServices,
    GpusAllRegions,
    HealthChecks,
    Images,
    Instances,
    InstanceGroups,
    InstanceGroupManagers,
    InstanceTemplates,
    Interconnects,
    InterconnectAttachmentsPerRegion,
    InterconnectAttachmentsTotalMbps,
    InterconnectTotalGbps,
    InternalAddresses,
    InternalTrafficDirectorForwardingRules,
    InPlaceSnapshots,
    InUseAddresses,
    InUseBackupSchedules,
    InUseSnapshotSchedules,
    LocalSsdTotalGb,
    M1Cpus,
    M2Cpus,
    M3Cpus,
    MachineImages,
    N2aCpus,
    N2dCpus,
    N2Cpus,
    Networks,
    NetworkAttachments,
    NetworkEndpointGroups,
    NetworkFirewallPolicies,
    NetLbSecurityPoliciesPerRegion,
    NetLbSecurityPolicyRulesPerRegion,
    NetLbSecurityPolicyRuleAttributesPerRegion,
    NodeGroups,
    NodeTemplates,
    NvidiaA10080gbGpus,
    NvidiaA100Gpus,
    NvidiaK80Gpus,
    NvidiaL4Gpus,
    NvidiaP100Gpus,
    NvidiaP100VwsGpus,
    NvidiaP4Gpus,
    NvidiaP4VwsGpus,
    NvidiaT4Gpus,
    NvidiaT4VwsGpus,
    NvidiaV100Gpus,
    PacketMirrorings,
    PdExtremeTotalProvisionedIops,
    PreemptibleCpus,
    PreemptibleLocalSsdGb,
    PreemptibleNvidiaA10080gbGpus,
    PreemptibleNvidiaA100Gpus,
    PreemptibleNvidiaK80Gpus,
    PreemptibleNvidiaL4Gpus,
    PreemptibleNvidiaP100Gpus,
    PreemptibleNvidiaP100VwsGpus,
    PreemptibleNvidiaP4Gpus,
    PreemptibleNvidiaP4VwsGpus,
    PreemptibleNvidiaT4Gpus,
    PreemptibleNvidiaT4VwsGpus,
    PreemptibleNvidiaV100Gpus,
    PscIlbConsumerForwardingRulesPerProducerNetwork,
    PscInternalLbForwardingRules,
    PublicAdvertisedPrefixes,
    PublicDelegatedPrefixes,
    RegionalAutoscalers,
    RegionalExternalManagedBackendServices,
    RegionalExternalNetworkLbBackendServices,
    RegionalInstanceGroupManagers,
    RegionalInternalLbBackendServices,
    RegionalInternalManagedBackendServices,
    Reservations,
    ResourcePolicies,
    Routers,
    Routes,
    SecurityPolicies,
    SecurityPoliciesPerRegion,
    SecurityPolicyAdvancedRulesPerRegion,
    SecurityPolicyCevalRules,
    SecurityPolicyRules,
    SecurityPolicyRulesPerRegion,
    ServiceAttachments,
    Snapshots,
    SsdTotalGb,
    SslCertificates,
    StaticAddresses,
    StaticByoipAddresses,
    StaticExternalIpv6AddressRanges,
    Subnetworks,
    T2aCpus,
    T2dCpus,
    TargetHttpsProxies,
    TargetHttpProxies,
    TargetInstances,
    TargetPools,
    TargetSslProxies,
    TargetTcpProxies,
    TargetVpnGateways,
    UrlMaps,
    VpnGateways,
    VpnTunnels,
    XpnServiceProjects,

}

[Output Only] Name of the quota metric.

Variants

UndefinedMetric

A value indicating that the enum field is not set.

A2Cpus

AffinityGroups

Autoscalers

BackendBuckets

BackendServices

C2dCpus

C2Cpus

C3Cpus

Commitments

CommittedA2Cpus

CommittedC2dCpus

CommittedC2Cpus

CommittedC3Cpus

CommittedCpus

CommittedE2Cpus

CommittedLicenses

CommittedLocalSsdTotalGb

CommittedM3Cpus

CommittedMemoryOptimizedCpus

CommittedN2aCpus

CommittedN2dCpus

CommittedN2Cpus

CommittedNvidiaA10080gbGpus

CommittedNvidiaA100Gpus

CommittedNvidiaK80Gpus

CommittedNvidiaL4Gpus

CommittedNvidiaP100Gpus

CommittedNvidiaP4Gpus

CommittedNvidiaT4Gpus

CommittedNvidiaV100Gpus

CommittedT2aCpus

CommittedT2dCpus

Cpus

Guest CPUs

CpusAllRegions

DisksTotalGb

E2Cpus

ExternalManagedForwardingRules

ExternalNetworkLbForwardingRules

ExternalProtocolForwardingRules

ExternalVpnGateways

Firewalls

ForwardingRules

GlobalExternalManagedBackendServices

GlobalExternalManagedForwardingRules

GlobalExternalProxyLbBackendServices

GlobalInternalAddresses

GlobalInternalManagedBackendServices

GlobalInternalTrafficDirectorBackendServices

GpusAllRegions

HealthChecks

Images

Instances

InstanceGroups

InstanceGroupManagers

InstanceTemplates

Interconnects

InterconnectAttachmentsPerRegion

InterconnectAttachmentsTotalMbps

InterconnectTotalGbps

InternalAddresses

InternalTrafficDirectorForwardingRules

InPlaceSnapshots

InUseAddresses

InUseBackupSchedules

InUseSnapshotSchedules

LocalSsdTotalGb

M1Cpus

M2Cpus

M3Cpus

MachineImages

N2aCpus

N2dCpus

N2Cpus

Networks

NetworkAttachments

NetworkEndpointGroups

NetworkFirewallPolicies

NetLbSecurityPoliciesPerRegion

NetLbSecurityPolicyRulesPerRegion

NetLbSecurityPolicyRuleAttributesPerRegion

NodeGroups

NodeTemplates

NvidiaA10080gbGpus

NvidiaA100Gpus

NvidiaK80Gpus

NvidiaL4Gpus

NvidiaP100Gpus

NvidiaP100VwsGpus

NvidiaP4Gpus

NvidiaP4VwsGpus

NvidiaT4Gpus

NvidiaT4VwsGpus

NvidiaV100Gpus

PacketMirrorings

PdExtremeTotalProvisionedIops

PreemptibleCpus

PreemptibleLocalSsdGb

PreemptibleNvidiaA10080gbGpus

PreemptibleNvidiaA100Gpus

PreemptibleNvidiaK80Gpus

PreemptibleNvidiaL4Gpus

PreemptibleNvidiaP100Gpus

PreemptibleNvidiaP100VwsGpus

PreemptibleNvidiaP4Gpus

PreemptibleNvidiaP4VwsGpus

PreemptibleNvidiaT4Gpus

PreemptibleNvidiaT4VwsGpus

PreemptibleNvidiaV100Gpus

PscIlbConsumerForwardingRulesPerProducerNetwork

PscInternalLbForwardingRules

PublicAdvertisedPrefixes

PublicDelegatedPrefixes

RegionalAutoscalers

RegionalExternalManagedBackendServices

RegionalExternalNetworkLbBackendServices

RegionalInstanceGroupManagers

RegionalInternalLbBackendServices

RegionalInternalManagedBackendServices

Reservations

ResourcePolicies

Routers

Routes

SecurityPolicies

SecurityPoliciesPerRegion

SecurityPolicyAdvancedRulesPerRegion

SecurityPolicyCevalRules

SecurityPolicyRules

SecurityPolicyRulesPerRegion

ServiceAttachments

Snapshots

The total number of snapshots allowed for a single project.

SsdTotalGb

SslCertificates

StaticAddresses

StaticByoipAddresses

StaticExternalIpv6AddressRanges

Subnetworks

T2aCpus

T2dCpus

TargetHttpsProxies

TargetHttpProxies

TargetInstances

TargetPools

TargetSslProxies

TargetTcpProxies

TargetVpnGateways

UrlMaps

VpnGateways

VpnTunnels

XpnServiceProjects

Implementations

impl Metric

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Metric.

pub fn from_i32(value: i32) -> Option<Metric>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Metric, or None if value is not a valid variant.

impl Metric

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.