Struct google_api_proto::google::cloud::compute::v1::region_instance_group_managers_client::RegionInstanceGroupManagersClient

pub struct RegionInstanceGroupManagersClient<T> { /* private fields */ }

The RegionInstanceGroupManagers API.

Implementations

impl<T> RegionInstanceGroupManagersClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> RegionInstanceGroupManagersClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn abandon_instances( &mut self, request: impl IntoRequest<AbandonInstancesRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified instances to be immediately removed from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn apply_updates_to_instances( &mut self, request: impl IntoRequest<ApplyUpdatesToInstancesRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Apply updates to selected instances the managed instance group.

pub async fn create_instances( &mut self, request: impl IntoRequest<CreateInstancesRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Creates instances with per-instance configurations in this regional managed instance group. Instances are created using the current instance template. The create instances operation is marked DONE if the createInstances request is successful. The underlying actions take additional time. You must separately verify the status of the creating or actions with the listmanagedinstances method.

pub async fn delete( &mut self, request: impl IntoRequest<DeleteRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Deletes the specified managed instance group and all of the instances in that group.

pub async fn delete_instances( &mut self, request: impl IntoRequest<DeleteInstancesRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified instances in the managed instance group to be immediately deleted. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. The deleteInstances operation is marked DONE if the deleteInstances request is successful. The underlying actions take additional time. You must separately verify the status of the deleting action with the listmanagedinstances method. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn delete_per_instance_configs( &mut self, request: impl IntoRequest<DeletePerInstanceConfigsRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Deletes selected per-instance configurations for the managed instance group.

pub async fn get( &mut self, request: impl IntoRequest<GetRegionInstanceGroupManagerRequest> ) -> Result<Response<InstanceGroupManager>, Status>

Returns all of the details about the specified managed instance group.

pub async fn insert( &mut self, request: impl IntoRequest<InsertRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Creates a managed instance group using the information that you specify in the request. After the group is created, instances in the group are created using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method. A regional managed instance group can contain up to 2000 instances.

pub async fn list( &mut self, request: impl IntoRequest<ListRegionInstanceGroupManagersRequest> ) -> Result<Response<RegionInstanceGroupManagerList>, Status>

Retrieves the list of managed instance groups that are contained within the specified region.

pub async fn list_errors( &mut self, request: impl IntoRequest<ListErrorsRegionInstanceGroupManagersRequest> ) -> Result<Response<RegionInstanceGroupManagersListErrorsResponse>, Status>

Lists all errors thrown by actions on instances for a given regional managed instance group. The filter and orderBy query parameters are not supported.

pub async fn list_managed_instances( &mut self, request: impl IntoRequest<ListManagedInstancesRegionInstanceGroupManagersRequest> ) -> Result<Response<RegionInstanceGroupManagersListInstancesResponse>, Status>

Lists the instances in the managed instance group and instances that are scheduled to be created. The list includes any current actions that the group has scheduled for its instances. The orderBy query parameter is not supported. The pageToken query parameter is supported only in the alpha and beta API and only if the group’s listManagedInstancesResults field is set to PAGINATED.

pub async fn list_per_instance_configs( &mut self, request: impl IntoRequest<ListPerInstanceConfigsRegionInstanceGroupManagersRequest> ) -> Result<Response<RegionInstanceGroupManagersListInstanceConfigsResp>, Status>

Lists all of the per-instance configurations defined for the managed instance group. The orderBy query parameter is not supported.

pub async fn patch( &mut self, request: impl IntoRequest<PatchRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Updates a managed instance group using the information that you specify in the request. This operation is marked as DONE when the group is patched even if the instances in the group are still in the process of being patched. You must separately verify the status of the individual instances with the listmanagedinstances method. This method supports PATCH semantics and uses the JSON merge patch format and processing rules. If you update your group to specify a new template or instance configuration, it’s possible that your intended specification for each VM in the group is different from the current state of that VM. To learn how to apply an updated configuration to the VMs in a MIG, see Updating instances in a MIG.

pub async fn patch_per_instance_configs( &mut self, request: impl IntoRequest<PatchPerInstanceConfigsRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Inserts or patches per-instance configurations for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.

pub async fn recreate_instances( &mut self, request: impl IntoRequest<RecreateInstancesRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Flags the specified VM instances in the managed instance group to be immediately recreated. Each instance is recreated using the group’s current configuration. This operation is marked as DONE when the flag is set even if the instances have not yet been recreated. You must separately verify the status of each instance by checking its currentAction field; for more information, see Checking the status of managed instances. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted. You can specify a maximum of 1000 instances with this method per request.

pub async fn resize( &mut self, request: impl IntoRequest<ResizeRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Changes the intended size of the managed instance group. If you increase the size, the group creates new instances using the current instance template. If you decrease the size, the group deletes one or more instances. The resize operation is marked DONE if the resize request is successful. The underlying actions take additional time. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method. If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

pub async fn set_instance_template( &mut self, request: impl IntoRequest<SetInstanceTemplateRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Sets the instance template to use when creating new instances or recreating instances in this group. Existing instances are not affected.

pub async fn set_target_pools( &mut self, request: impl IntoRequest<SetTargetPoolsRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.

pub async fn update_per_instance_configs( &mut self, request: impl IntoRequest<UpdatePerInstanceConfigsRegionInstanceGroupManagerRequest> ) -> Result<Response<Operation>, Status>

Inserts or updates per-instance configurations for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.