Struct google_api_proto::google::cloud::compute::v1::AttachedDiskInitializeParams

pub struct AttachedDiskInitializeParams {

    pub architecture: Option<String>,
    pub description: Option<String>,
    pub disk_name: Option<String>,
    pub disk_size_gb: Option<i64>,
    pub disk_type: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub licenses: Vec<String>,
    pub on_update_action: Option<String>,
    pub provisioned_iops: Option<i64>,
    pub provisioned_throughput: Option<i64>,
    pub replica_zones: Vec<String>,
    pub resource_manager_tags: BTreeMap<String, String>,
    pub resource_policies: Vec<String>,
    pub source_image: Option<String>,
    pub source_image_encryption_key: Option<CustomerEncryptionKey>,
    pub source_snapshot: Option<String>,
    pub source_snapshot_encryption_key: Option<CustomerEncryptionKey>,

}

[Input Only] Specifies the parameters for a new disk that will be created alongside the new instance. Use initialization parameters to create boot disks or local SSDs attached to the new instance. This field is persisted and returned for instanceTemplate and not returned in the context of instance. This property is mutually exclusive with the source property; you can only define one or the other, but not both.

Fields
architecture: Option<String>

The architecture of the attached disk. Valid values are arm64 or x86_64. Check the Architecture enum for the list of possible values.

description: Option<String>

An optional description. Provide this property when creating the disk.

disk_name: Option<String>

Specifies the disk name. If not specified, the default is to use the name of the instance. If a disk with the same name already exists in the given region, the existing disk is attached to the new instance and the new disk is not created.

disk_size_gb: Option<i64>

Specifies the size of the disk in base-2 GB. The size must be at least 10 GB. If you specify a sourceImage, which is required for boot disks, the default size is the size of the sourceImage. If you do not specify a sourceImage, the default disk size is 500 GB.

disk_type: Option<String>

Specifies the disk type to use to create the instance. If not specified, the default is pd-standard, specified using the full URL. For example: https://www.googleapis.com/compute/v1/projects/project/zones/zone /diskTypes/pd-standard For a full list of acceptable values, see Persistent disk types. If you specify this field when creating a VM, you can provide either the full or partial URL. For example, the following values are valid: - https://www.googleapis.com/compute/v1/projects/project/zones/zone /diskTypes/diskType - projects/project/zones/zone/diskTypes/diskType - zones/zone/diskTypes/diskType If you specify this field when creating or updating an instance template or all-instances configuration, specify the type of the disk, not the URL. For example: pd-standard.

labels: BTreeMap<String, String>

Labels to apply to this disk. These can be later modified by the disks.setLabels method. This field is only applicable for persistent disks.

licenses: Vec<String>

A list of publicly visible licenses. Reserved for Google’s use.

on_update_action: Option<String>

Specifies which action to take on instance update with this disk. Default is to use the existing disk. Check the OnUpdateAction enum for the list of possible values.

provisioned_iops: Option<i64>

Indicates how many IOPS to provision for the disk. This sets the number of I/O operations per second that the disk can handle. Values must be between 10,000 and 120,000. For more details, see the Extreme persistent disk documentation.

provisioned_throughput: Option<i64>

Indicates how much throughput to provision for the disk. This sets the number of throughput mb per second that the disk can handle. Values must be between 1 and 7,124.

replica_zones: Vec<String>

Required for each regional disk associated with the instance. Specify the URLs of the zones where the disk should be replicated to. You must provide exactly two replica zones, and one zone must be the same as the instance zone.

resource_manager_tags: BTreeMap<String, String>

Resource manager tags to be bound to the disk. Tag keys and values have the same definition as resource manager tags. Keys must be in the format tagKeys/{tag_key_id}, and values are in the format tagValues/456. The field is ignored (both PUT & PATCH) when empty.

resource_policies: Vec<String>

Resource policies applied to this disk for automatic snapshot creations. Specified using the full or partial URL. For instance template, specify only the resource policy name.

source_image: Option<String>

The source image to create this disk. When creating a new instance, one of initializeParams.sourceImage or initializeParams.sourceSnapshot or disks.source is required except for local SSD. To create a disk with one of the public operating system images, specify the image by its family name. For example, specify family/debian-9 to use the latest Debian 9 image: projects/debian-cloud/global/images/family/debian-9 Alternatively, use a specific version of a public operating system image: projects/debian-cloud/global/images/debian-9-stretch-vYYYYMMDD To create a disk with a custom image that you created, specify the image name in the following format: global/images/my-custom-image You can also specify a custom image by its image family, which returns the latest version of the image in that family. Replace the image name with family/family-name: global/images/family/my-image-family If the source image is deleted later, this field will not be set.

source_image_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source image. Required if the source image is protected by a customer-supplied encryption key. InstanceTemplate and InstancePropertiesPatch do not store customer-supplied encryption keys, so you cannot create disks for instances in a managed instance group if the source images are encrypted with your own keys.

source_snapshot: Option<String>

The source snapshot to create this disk. When creating a new instance, one of initializeParams.sourceSnapshot or initializeParams.sourceImage or disks.source is required except for local SSD. To create a disk with a snapshot that you created, specify the snapshot name in the following format: global/snapshots/my-backup If the source snapshot is deleted later, this field will not be set.

source_snapshot_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source snapshot.

Implementations

impl AttachedDiskInitializeParams

pub fn source_image(&self) -> &str

Returns the value of source_image, or the default value if source_image is unset.

pub fn disk_name(&self) -> &str

Returns the value of disk_name, or the default value if disk_name is unset.

pub fn disk_type(&self) -> &str

Returns the value of disk_type, or the default value if disk_type is unset.

pub fn source_snapshot(&self) -> &str

Returns the value of source_snapshot, or the default value if source_snapshot is unset.

pub fn provisioned_iops(&self) -> i64

Returns the value of provisioned_iops, or the default value if provisioned_iops is unset.

pub fn on_update_action(&self) -> &str

Returns the value of on_update_action, or the default value if on_update_action is unset.

pub fn architecture(&self) -> &str

Returns the value of architecture, or the default value if architecture is unset.

pub fn disk_size_gb(&self) -> i64

Returns the value of disk_size_gb, or the default value if disk_size_gb is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn provisioned_throughput(&self) -> i64

Returns the value of provisioned_throughput, or the default value if provisioned_throughput is unset.