Struct google_api_proto::google::cloud::compute::v1::Disk

pub struct Disk {

    pub architecture: Option<String>,
    pub async_primary_disk: Option<DiskAsyncReplication>,
    pub async_secondary_disks: BTreeMap<String, DiskAsyncReplicationList>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub disk_encryption_key: Option<CustomerEncryptionKey>,
    pub enable_confidential_compute: Option<bool>,
    pub guest_os_features: Vec<GuestOsFeature>,
    pub id: Option<u64>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub last_attach_timestamp: Option<String>,
    pub last_detach_timestamp: Option<String>,
    pub license_codes: Vec<i64>,
    pub licenses: Vec<String>,
    pub location_hint: Option<String>,
    pub name: Option<String>,
    pub options: Option<String>,
    pub params: Option<DiskParams>,
    pub physical_block_size_bytes: Option<i64>,
    pub provisioned_iops: Option<i64>,
    pub provisioned_throughput: Option<i64>,
    pub region: Option<String>,
    pub replica_zones: Vec<String>,
    pub resource_policies: Vec<String>,
    pub resource_status: Option<DiskResourceStatus>,
    pub satisfies_pzi: Option<bool>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub size_gb: Option<i64>,
    pub source_consistency_group_policy: Option<String>,
    pub source_consistency_group_policy_id: Option<String>,
    pub source_disk: Option<String>,
    pub source_disk_id: Option<String>,
    pub source_image: Option<String>,
    pub source_image_encryption_key: Option<CustomerEncryptionKey>,
    pub source_image_id: Option<String>,
    pub source_snapshot: Option<String>,
    pub source_snapshot_encryption_key: Option<CustomerEncryptionKey>,
    pub source_snapshot_id: Option<String>,
    pub source_storage_object: Option<String>,
    pub status: Option<String>,
    pub type: Option<String>,
    pub users: Vec<String>,
    pub zone: Option<String>,

}

Represents a Persistent Disk resource. Google Compute Engine has two Disk resources: * Zonal * Regional Persistent disks are required for running your VM instances. Create both boot and non-boot (data) persistent disks. For more information, read Persistent Disks. For more storage options, read Storage options. The disks resource represents a zonal persistent disk. For more information, read Zonal persistent disks. The regionDisks resource represents a regional persistent disk. For more information, read Regional resources.

Fields
architecture: Option<String>

The architecture of the disk. Valid values are ARM64 or X86_64. Check the Architecture enum for the list of possible values.

async_primary_disk: Option<DiskAsyncReplication>

Disk asynchronously replicated into this disk.

async_secondary_disks: BTreeMap<String, DiskAsyncReplicationList>

[Output Only] A list of disks this disk is asynchronously replicated to.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disk_encryption_key: Option<CustomerEncryptionKey>

Encrypts the disk using a customer-supplied encryption key or a customer-managed encryption key. Encryption keys do not protect access to metadata of the disk. After you encrypt a disk with a customer-supplied key, you must provide the same key if you use the disk later. For example, to create a disk snapshot, to create a disk image, to create a machine image, or to attach the disk to a virtual machine. After you encrypt a disk with a customer-managed key, the diskEncryptionKey.kmsKeyName is set to a key version name once the disk is created. The disk is encrypted with this version of the key. In the response, diskEncryptionKey.kmsKeyName appears in the following format: “diskEncryptionKey.kmsKeyName”: “projects/kms_project_id/locations/region/keyRings/ key_region/cryptoKeys/key /cryptoKeysVersions/version If you do not provide an encryption key when creating the disk, then the disk is encrypted using an automatically generated key and you don’t need to provide a key to use the disk later.

enable_confidential_compute: Option<bool>

Whether this disk is using confidential compute mode.

guest_os_features: Vec<GuestOsFeature>

A list of features to enable on the guest operating system. Applicable only for bootable images. Read Enabling guest operating system features to see a list of available options.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#disk for disks.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this disk, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a disk.

labels: BTreeMap<String, String>

Labels to apply to this disk. These can be later modified by the setLabels method.

last_attach_timestamp: Option<String>

[Output Only] Last attach timestamp in RFC3339 text format.

last_detach_timestamp: Option<String>

[Output Only] Last detach timestamp in RFC3339 text format.

license_codes: Vec<i64>

Integer license codes indicating which licenses are attached to this disk.

licenses: Vec<String>

A list of publicly visible licenses. Reserved for Google’s use.

location_hint: Option<String>

An opaque location hint used to place the disk close to other resources. This field is for use by internal tools that use the public API.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

options: Option<String>

Internal use only.

params: Option<DiskParams>

Input only. [Input Only] Additional params passed with the request, but not persisted as part of resource payload.

physical_block_size_bytes: Option<i64>

Physical block size of the persistent disk, in bytes. If not present in a request, a default value is used. The currently supported size is 4096, other sizes may be added in the future. If an unsupported value is requested, the error message will list the supported values for the caller’s project.

provisioned_iops: Option<i64>

Indicates how many IOPS to provision for the disk. This sets the number of I/O operations per second that the disk can handle. Values must be between 10,000 and 120,000. For more details, see the Extreme persistent disk documentation.

provisioned_throughput: Option<i64>

Indicates how much throughput to provision for the disk. This sets the number of throughput mb per second that the disk can handle. Values must be between 1 and 7,124.

region: Option<String>

[Output Only] URL of the region where the disk resides. Only applicable for regional resources. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

replica_zones: Vec<String>

URLs of the zones where the disk should be replicated to. Only applicable for regional resources.

resource_policies: Vec<String>

Resource policies applied to this disk for automatic snapshot creations.

resource_status: Option<DiskResourceStatus>

[Output Only] Status information for the disk resource.

satisfies_pzi: Option<bool>

Output only. Reserved for future use.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

self_link: Option<String>

[Output Only] Server-defined fully-qualified URL for this resource.

size_gb: Option<i64>

Size, in GB, of the persistent disk. You can specify this field when creating a persistent disk using the sourceImage, sourceSnapshot, or sourceDisk parameter, or specify it alone to create an empty persistent disk. If you specify this field along with a source, the value of sizeGb must not be less than the size of the source. Acceptable values are greater than 0.

source_consistency_group_policy: Option<String>

[Output Only] URL of the DiskConsistencyGroupPolicy for a secondary disk that was created using a consistency group.

source_consistency_group_policy_id: Option<String>

[Output Only] ID of the DiskConsistencyGroupPolicy for a secondary disk that was created using a consistency group.

source_disk: Option<String>

The source disk used to create this disk. You can provide this as a partial or full URL to the resource. For example, the following are valid values: - https://www.googleapis.com/compute/v1/projects/project/zones/zone /disks/disk - https://www.googleapis.com/compute/v1/projects/project/regions/region /disks/disk - projects/project/zones/zone/disks/disk - projects/project/regions/region/disks/disk - zones/zone/disks/disk - regions/region/disks/disk

source_disk_id: Option<String>

[Output Only] The unique ID of the disk used to create this disk. This value identifies the exact disk that was used to create this persistent disk. For example, if you created the persistent disk from a disk that was later deleted and recreated under the same name, the source disk ID would identify the exact version of the disk that was used.

source_image: Option<String>

The source image used to create this disk. If the source image is deleted, this field will not be set. To create a disk with one of the public operating system images, specify the image by its family name. For example, specify family/debian-9 to use the latest Debian 9 image: projects/debian-cloud/global/images/family/debian-9 Alternatively, use a specific version of a public operating system image: projects/debian-cloud/global/images/debian-9-stretch-vYYYYMMDD To create a disk with a custom image that you created, specify the image name in the following format: global/images/my-custom-image You can also specify a custom image by its image family, which returns the latest version of the image in that family. Replace the image name with family/family-name: global/images/family/my-image-family

source_image_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source image. Required if the source image is protected by a customer-supplied encryption key.

source_image_id: Option<String>

[Output Only] The ID value of the image used to create this disk. This value identifies the exact image that was used to create this persistent disk. For example, if you created the persistent disk from an image that was later deleted and recreated under the same name, the source image ID would identify the exact version of the image that was used.

source_snapshot: Option<String>

The source snapshot used to create this disk. You can provide this as a partial or full URL to the resource. For example, the following are valid values: - https://www.googleapis.com/compute/v1/projects/project /global/snapshots/snapshot - projects/project/global/snapshots/snapshot - global/snapshots/snapshot

source_snapshot_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source snapshot. Required if the source snapshot is protected by a customer-supplied encryption key.

source_snapshot_id: Option<String>

[Output Only] The unique ID of the snapshot used to create this disk. This value identifies the exact snapshot that was used to create this persistent disk. For example, if you created the persistent disk from a snapshot that was later deleted and recreated under the same name, the source snapshot ID would identify the exact version of the snapshot that was used.

source_storage_object: Option<String>

The full Google Cloud Storage URI where the disk image is stored. This file must be a gzip-compressed tarball whose name ends in .tar.gz or virtual machine disk whose name ends in vmdk. Valid URIs may start with gs:// or https://storage.googleapis.com/. This flag is not optimized for creating multiple disks from a source storage object. To create many disks from a source storage object, use gcloud compute images import instead.

status: Option<String>

[Output Only] The status of disk creation. - CREATING: Disk is provisioning. - RESTORING: Source data is being copied into the disk. - FAILED: Disk creation failed. - READY: Disk is ready for use. - DELETING: Disk is deleting. Check the Status enum for the list of possible values.

type: Option<String>

URL of the disk type resource describing which disk type to use to create the disk. Provide this when creating the disk. For example: projects/project /zones/zone/diskTypes/pd-ssd . See Persistent disk types.

users: Vec<String>

[Output Only] Links to the users of the disk (attached instances) in form: projects/project/zones/zone/instances/instance

zone: Option<String>

[Output Only] URL of the zone where the disk resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

Implementations

impl Disk

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn type(&self) -> &str

Returns the value of type, or the default value if type is unset.

pub fn zone(&self) -> &str

Returns the value of zone, or the default value if zone is unset.

pub fn source_consistency_group_policy(&self) -> &str

Returns the value of source_consistency_group_policy, or the default value if source_consistency_group_policy is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn last_attach_timestamp(&self) -> &str

Returns the value of last_attach_timestamp, or the default value if last_attach_timestamp is unset.

pub fn source_image(&self) -> &str

Returns the value of source_image, or the default value if source_image is unset.

pub fn source_image_id(&self) -> &str

Returns the value of source_image_id, or the default value if source_image_id is unset.

pub fn last_detach_timestamp(&self) -> &str

Returns the value of last_detach_timestamp, or the default value if last_detach_timestamp is unset.

pub fn source_snapshot_id(&self) -> &str

Returns the value of source_snapshot_id, or the default value if source_snapshot_id is unset.

pub fn enable_confidential_compute(&self) -> bool

Returns the value of enable_confidential_compute, or the default value if enable_confidential_compute is unset.

pub fn source_snapshot(&self) -> &str

Returns the value of source_snapshot, or the default value if source_snapshot is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn provisioned_iops(&self) -> i64

Returns the value of provisioned_iops, or the default value if provisioned_iops is unset.

pub fn source_storage_object(&self) -> &str

Returns the value of source_storage_object, or the default value if source_storage_object is unset.

pub fn source_consistency_group_policy_id(&self) -> &str

Returns the value of source_consistency_group_policy_id, or the default value if source_consistency_group_policy_id is unset.

pub fn architecture(&self) -> &str

Returns the value of architecture, or the default value if architecture is unset.

pub fn location_hint(&self) -> &str

Returns the value of location_hint, or the default value if location_hint is unset.

pub fn options(&self) -> &str

Returns the value of options, or the default value if options is unset.

pub fn physical_block_size_bytes(&self) -> i64

Returns the value of physical_block_size_bytes, or the default value if physical_block_size_bytes is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn source_disk(&self) -> &str

Returns the value of source_disk, or the default value if source_disk is unset.

pub fn source_disk_id(&self) -> &str

Returns the value of source_disk_id, or the default value if source_disk_id is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn satisfies_pzi(&self) -> bool

Returns the value of satisfies_pzi, or the default value if satisfies_pzi is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.

pub fn size_gb(&self) -> i64

Returns the value of size_gb, or the default value if size_gb is unset.

pub fn provisioned_throughput(&self) -> i64

Returns the value of provisioned_throughput, or the default value if provisioned_throughput is unset.