Struct google_api_proto::google::cloud::compute::v1::ForwardingRule

pub struct ForwardingRule {

    pub i_p_address: Option<String>,
    pub i_p_protocol: Option<String>,
    pub all_ports: Option<bool>,
    pub allow_global_access: Option<bool>,
    pub allow_psc_global_access: Option<bool>,
    pub backend_service: Option<String>,
    pub base_forwarding_rule: Option<String>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub fingerprint: Option<String>,
    pub id: Option<u64>,
    pub ip_version: Option<String>,
    pub is_mirroring_collector: Option<bool>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub load_balancing_scheme: Option<String>,
    pub metadata_filters: Vec<MetadataFilter>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub network_tier: Option<String>,
    pub no_automate_dns_zone: Option<bool>,
    pub port_range: Option<String>,
    pub ports: Vec<String>,
    pub psc_connection_id: Option<u64>,
    pub psc_connection_status: Option<String>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub service_directory_registrations: Vec<ForwardingRuleServiceDirectoryRegistration>,
    pub service_label: Option<String>,
    pub service_name: Option<String>,
    pub source_ip_ranges: Vec<String>,
    pub subnetwork: Option<String>,
    pub target: Option<String>,

}

Represents a Forwarding Rule resource. Forwarding rule resources in Google Cloud can be either regional or global in scope: * Global * Regional A forwarding rule and its corresponding IP address represent the frontend configuration of a Google Cloud load balancer. Forwarding rules can also reference target instances and Cloud VPN Classic gateways (targetVpnGateway). For more information, read Forwarding rule concepts and Using protocol forwarding.

Fields
i_p_address: Option<String>

IP address for which this forwarding rule accepts traffic. When a client sends traffic to this IP address, the forwarding rule directs the traffic to the referenced target or backendService. While creating a forwarding rule, specifying an IPAddress is required under the following circumstances: - When the target is set to targetGrpcProxy and validateForProxyless is set to true, the IPAddress should be set to 0.0.0.0. - When the target is a Private Service Connect Google APIs bundle, you must specify an IPAddress. Otherwise, you can optionally specify an IP address that references an existing static (reserved) IP address resource. When omitted, Google Cloud assigns an ephemeral IP address. Use one of the following formats to specify an IP address while creating a forwarding rule: * IP address number, as in 100.1.2.3 * IPv6 address range, as in 2600:1234::/96 * Full resource URL, as in https://www.googleapis.com/compute/v1/projects/ project_id/regions/region/addresses/address-name * Partial URL or by name, as in: - projects/project_id/regions/region/addresses/address-name - regions/region/addresses/address-name - global/addresses/address-name - address-name The forwarding rule’s target or backendService, and in most cases, also the loadBalancingScheme, determine the type of IP address that you can use. For detailed information, see IP address specifications. When reading an IPAddress, the API always returns the IP address number.

i_p_protocol: Option<String>

The IP protocol to which this rule applies. For protocol forwarding, valid options are TCP, UDP, ESP, AH, SCTP, ICMP and L3_DEFAULT. The valid IP protocols are different for different load balancing products as described in Load balancing features. Check the IPProtocolEnum enum for the list of possible values.

all_ports: Option<bool>

The ports, portRange, and allPorts fields are mutually exclusive. Only packets addressed to ports in the specified range will be forwarded to the backends configured with this forwarding rule. The allPorts field has the following limitations: - It requires that the forwarding rule IPProtocol be TCP, UDP, SCTP, or L3_DEFAULT. - It’s applicable only to the following products: internal passthrough Network Load Balancers, backend service-based external passthrough Network Load Balancers, and internal and external protocol forwarding. - Set this field to true to allow packets addressed to any port or packets lacking destination port information (for example, UDP fragments after the first fragment) to be forwarded to the backends configured with this forwarding rule. The L3_DEFAULT protocol requires allPorts be set to true.

allow_global_access: Option<bool>

If set to true, clients can access the internal passthrough Network Load Balancers, the regional internal Application Load Balancer, and the regional internal proxy Network Load Balancer from all regions. If false, only allows access from the local region the load balancer is located at. Note that for INTERNAL_MANAGED forwarding rules, this field cannot be changed after the forwarding rule is created.

allow_psc_global_access: Option<bool>

This is used in PSC consumer ForwardingRule to control whether the PSC endpoint can be accessed from another region.

backend_service: Option<String>

Identifies the backend service to which the forwarding rule sends traffic. Required for internal and external passthrough Network Load Balancers; must be omitted for all other load balancer types.

base_forwarding_rule: Option<String>

[Output Only] The URL for the corresponding base forwarding rule. By base forwarding rule, we mean the forwarding rule that has the same IP address, protocol, and port settings with the current forwarding rule, but without sourceIPRanges specified. Always empty if the current forwarding rule does not have sourceIPRanges specified.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a ForwardingRule. Include the fingerprint in patch request to ensure that you do not overwrite changes that were applied from another concurrent request. To see the latest fingerprint, make a get() request to retrieve a ForwardingRule.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

ip_version: Option<String>

The IP Version that will be used by this forwarding rule. Valid options are IPV4 or IPV6. Check the IpVersion enum for the list of possible values.

is_mirroring_collector: Option<bool>

Indicates whether or not this load balancer can be used as a collector for packet mirroring. To prevent mirroring loops, instances behind this load balancer will not have their traffic mirrored even if a PacketMirroring rule applies to them. This can only be set to true for load balancers that have their loadBalancingScheme set to INTERNAL.

kind: Option<String>

[Output Only] Type of the resource. Always compute#forwardingRule for forwarding rule resources.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this resource, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a ForwardingRule.

labels: BTreeMap<String, String>

Labels for this resource. These can only be added or modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.

load_balancing_scheme: Option<String>

Specifies the forwarding rule type. For more information about forwarding rules, refer to Forwarding rule concepts. Check the LoadBalancingScheme enum for the list of possible values.

metadata_filters: Vec<MetadataFilter>

Opaque filter criteria used by load balancer to restrict routing configuration to a limited set of xDS compliant clients. In their xDS requests to load balancer, xDS clients present node metadata. When there is a match, the relevant configuration is made available to those proxies. Otherwise, all the resources (e.g. TargetHttpProxy, UrlMap) referenced by the ForwardingRule are not visible to those proxies. For each metadataFilter in this list, if its filterMatchCriteria is set to MATCH_ANY, at least one of the filterLabels must match the corresponding label provided in the metadata. If its filterMatchCriteria is set to MATCH_ALL, then all of its filterLabels must match with corresponding labels provided in the metadata. If multiple metadataFilters are specified, all of them need to be satisfied in order to be considered a match. metadataFilters specified here will be applifed before those specified in the UrlMap that this ForwardingRule references. metadataFilters only applies to Loadbalancers that have their loadBalancingScheme set to INTERNAL_SELF_MANAGED.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash. For Private Service Connect forwarding rules that forward traffic to Google APIs, the forwarding rule name must be a 1-20 characters string with lowercase letters and numbers and must start with a letter.

network: Option<String>

This field is not used for global external load balancing. For internal passthrough Network Load Balancers, this field identifies the network that the load balanced IP should belong to for this forwarding rule. If the subnetwork is specified, the network of the subnetwork will be used. If neither subnetwork nor this field is specified, the default network will be used. For Private Service Connect forwarding rules that forward traffic to Google APIs, a network must be provided.

network_tier: Option<String>

This signifies the networking tier used for configuring this load balancer and can only take the following values: PREMIUM, STANDARD. For regional ForwardingRule, the valid values are PREMIUM and STANDARD. For GlobalForwardingRule, the valid value is PREMIUM. If this field is not specified, it is assumed to be PREMIUM. If IPAddress is specified, this value must be equal to the networkTier of the Address. Check the NetworkTier enum for the list of possible values.

no_automate_dns_zone: Option<bool>

This is used in PSC consumer ForwardingRule to control whether it should try to auto-generate a DNS zone or not. Non-PSC forwarding rules do not use this field. Once set, this field is not mutable.

port_range: Option<String>

The ports, portRange, and allPorts fields are mutually exclusive. Only packets addressed to ports in the specified range will be forwarded to the backends configured with this forwarding rule. The portRange field has the following limitations: - It requires that the forwarding rule IPProtocol be TCP, UDP, or SCTP, and - It’s applicable only to the following products: external passthrough Network Load Balancers, internal and external proxy Network Load Balancers, internal and external Application Load Balancers, external protocol forwarding, and Classic VPN. - Some products have restrictions on what ports can be used. See port specifications for details. For external forwarding rules, two or more forwarding rules cannot use the same [IPAddress, IPProtocol] pair, and cannot have overlapping portRanges. For internal forwarding rules within the same VPC network, two or more forwarding rules cannot use the same [IPAddress, IPProtocol] pair, and cannot have overlapping portRanges. @pattern: \d+(?:-\d+)?

ports: Vec<String>

The ports, portRange, and allPorts fields are mutually exclusive. Only packets addressed to ports in the specified range will be forwarded to the backends configured with this forwarding rule. The ports field has the following limitations: - It requires that the forwarding rule IPProtocol be TCP, UDP, or SCTP, and - It’s applicable only to the following products: internal passthrough Network Load Balancers, backend service-based external passthrough Network Load Balancers, and internal protocol forwarding. - You can specify a list of up to five ports by number, separated by commas. The ports can be contiguous or discontiguous. For external forwarding rules, two or more forwarding rules cannot use the same [IPAddress, IPProtocol] pair if they share at least one port number. For internal forwarding rules within the same VPC network, two or more forwarding rules cannot use the same [IPAddress, IPProtocol] pair if they share at least one port number. @pattern: \d+(?:-\d+)?

psc_connection_id: Option<u64>

[Output Only] The PSC connection id of the PSC forwarding rule.

psc_connection_status: Option<String>

Check the PscConnectionStatus enum for the list of possible values.

region: Option<String>

[Output Only] URL of the region where the regional forwarding rule resides. This field is not applicable to global forwarding rules. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

service_directory_registrations: Vec<ForwardingRuleServiceDirectoryRegistration>

Service Directory resources to register this forwarding rule with. Currently, only supports a single Service Directory resource.

service_label: Option<String>

An optional prefix to the service name for this forwarding rule. If specified, the prefix is the first label of the fully qualified service name. The label must be 1-63 characters long, and comply with RFC1035. Specifically, the label must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash. This field is only used for internal load balancing.

service_name: Option<String>

[Output Only] The internal fully qualified service name for this forwarding rule. This field is only used for internal load balancing.

source_ip_ranges: Vec<String>

If not empty, this forwarding rule will only forward the traffic when the source IP address matches one of the IP addresses or CIDR ranges set here. Note that a forwarding rule can only have up to 64 source IP ranges, and this field can only be used with a regional forwarding rule whose scheme is EXTERNAL. Each source_ip_range entry should be either an IP address (for example, 1.2.3.4) or a CIDR range (for example, 1.2.3.0/24).

subnetwork: Option<String>

This field identifies the subnetwork that the load balanced IP should belong to for this forwarding rule, used with internal load balancers and external passthrough Network Load Balancers with IPv6. If the network specified is in auto subnet mode, this field is optional. However, a subnetwork must be specified if the network is in custom subnet mode or when creating external forwarding rule with IPv6.

target: Option<String>

The URL of the target resource to receive the matched traffic. For regional forwarding rules, this target must be in the same region as the forwarding rule. For global forwarding rules, this target must be a global load balancing resource. The forwarded traffic must be of a type appropriate to the target object. - For load balancers, see the “Target” column in Port specifications. - For Private Service Connect forwarding rules that forward traffic to Google APIs, provide the name of a supported Google API bundle: - vpc-sc - APIs that support VPC Service Controls. - all-apis - All supported Google APIs. - For Private Service Connect forwarding rules that forward traffic to managed services, the target must be a service attachment. The target is not mutable once set as a service attachment.

Implementations

impl ForwardingRule

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn i_p_address(&self) -> &str

Returns the value of i_p_address, or the default value if i_p_address is unset.

pub fn no_automate_dns_zone(&self) -> bool

Returns the value of no_automate_dns_zone, or the default value if no_automate_dns_zone is unset.

pub fn is_mirroring_collector(&self) -> bool

Returns the value of is_mirroring_collector, or the default value if is_mirroring_collector is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn psc_connection_status(&self) -> &str

Returns the value of psc_connection_status, or the default value if psc_connection_status is unset.

pub fn target(&self) -> &str

Returns the value of target, or the default value if target is unset.

pub fn port_range(&self) -> &str

Returns the value of port_range, or the default value if port_range is unset.

pub fn network(&self) -> &str

Returns the value of network, or the default value if network is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn allow_psc_global_access(&self) -> bool

Returns the value of allow_psc_global_access, or the default value if allow_psc_global_access is unset.

pub fn psc_connection_id(&self) -> u64

Returns the value of psc_connection_id, or the default value if psc_connection_id is unset.

pub fn ip_version(&self) -> &str

Returns the value of ip_version, or the default value if ip_version is unset.

pub fn backend_service(&self) -> &str

Returns the value of backend_service, or the default value if backend_service is unset.

pub fn subnetwork(&self) -> &str

Returns the value of subnetwork, or the default value if subnetwork is unset.

pub fn service_name(&self) -> &str

Returns the value of service_name, or the default value if service_name is unset.

pub fn load_balancing_scheme(&self) -> &str

Returns the value of load_balancing_scheme, or the default value if load_balancing_scheme is unset.

pub fn service_label(&self) -> &str

Returns the value of service_label, or the default value if service_label is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn all_ports(&self) -> bool

Returns the value of all_ports, or the default value if all_ports is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn i_p_protocol(&self) -> &str

Returns the value of i_p_protocol, or the default value if i_p_protocol is unset.

pub fn allow_global_access(&self) -> bool

Returns the value of allow_global_access, or the default value if allow_global_access is unset.

pub fn network_tier(&self) -> &str

Returns the value of network_tier, or the default value if network_tier is unset.

pub fn base_forwarding_rule(&self) -> &str

Returns the value of base_forwarding_rule, or the default value if base_forwarding_rule is unset.