Struct google_api_proto::google::cloud::compute::v1::Interconnect

pub struct Interconnect {

    pub admin_enabled: Option<bool>,
    pub circuit_infos: Vec<InterconnectCircuitInfo>,
    pub creation_timestamp: Option<String>,
    pub customer_name: Option<String>,
    pub description: Option<String>,
    pub expected_outages: Vec<InterconnectOutageNotification>,
    pub google_ip_address: Option<String>,
    pub google_reference_id: Option<String>,
    pub id: Option<u64>,
    pub interconnect_attachments: Vec<String>,
    pub interconnect_type: Option<String>,
    pub kind: Option<String>,
    pub link_type: Option<String>,
    pub location: Option<String>,
    pub name: Option<String>,
    pub noc_contact_email: Option<String>,
    pub operational_status: Option<String>,
    pub peer_ip_address: Option<String>,
    pub provisioned_link_count: Option<i32>,
    pub requested_link_count: Option<i32>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub state: Option<String>,

}

Represents an Interconnect resource. An Interconnect resource is a dedicated connection between the GCP network and your on-premises network. For more information, read the Dedicated Interconnect Overview.

Fields
admin_enabled: Option<bool>

Administrative status of the interconnect. When this is set to true, the Interconnect is functional and can carry traffic. When set to false, no packets can be carried over the interconnect and no BGP routes are exchanged over it. By default, the status is set to true.

circuit_infos: Vec<InterconnectCircuitInfo>

[Output Only] A list of CircuitInfo objects, that describe the individual circuits in this LAG.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

customer_name: Option<String>

Customer name, to put in the Letter of Authorization as the party authorized to request a crossconnect.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

expected_outages: Vec<InterconnectOutageNotification>

[Output Only] A list of outages expected for this Interconnect.

google_ip_address: Option<String>

[Output Only] IP address configured on the Google side of the Interconnect link. This can be used only for ping tests.

google_reference_id: Option<String>

[Output Only] Google reference ID to be used when raising support tickets with Google or otherwise to debug backend connectivity issues.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

interconnect_attachments: Vec<String>

[Output Only] A list of the URLs of all InterconnectAttachments configured to use this Interconnect.

interconnect_type: Option<String>

Type of interconnect, which can take one of the following values: - PARTNER: A partner-managed interconnection shared between customers though a partner. - DEDICATED: A dedicated physical interconnection with the customer. Note that a value IT_PRIVATE has been deprecated in favor of DEDICATED. Check the InterconnectType enum for the list of possible values.

kind: Option<String>

[Output Only] Type of the resource. Always compute#interconnect for interconnects.

link_type: Option<String>

Type of link requested, which can take one of the following values: - LINK_TYPE_ETHERNET_10G_LR: A 10G Ethernet with LR optics - LINK_TYPE_ETHERNET_100G_LR: A 100G Ethernet with LR optics. Note that this field indicates the speed of each of the links in the bundle, not the speed of the entire bundle. Check the LinkType enum for the list of possible values.

location: Option<String>

URL of the InterconnectLocation object that represents where this connection is to be provisioned.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression \[a-z]([-a-z0-9]*[a-z0-9\])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

noc_contact_email: Option<String>

Email address to contact the customer NOC for operations and maintenance notifications regarding this Interconnect. If specified, this will be used for notifications in addition to all other forms described, such as Stackdriver logs alerting and Cloud Notifications.

operational_status: Option<String>

[Output Only] The current status of this Interconnect’s functionality, which can take one of the following values: - OS_ACTIVE: A valid Interconnect, which is turned up and is ready to use. Attachments may be provisioned on this Interconnect. - OS_UNPROVISIONED: An Interconnect that has not completed turnup. No attachments may be provisioned on this Interconnect. - OS_UNDER_MAINTENANCE: An Interconnect that is undergoing internal maintenance. No attachments may be provisioned or updated on this Interconnect. Check the OperationalStatus enum for the list of possible values.

peer_ip_address: Option<String>

[Output Only] IP address configured on the customer side of the Interconnect link. The customer should configure this IP address during turnup when prompted by Google NOC. This can be used only for ping tests.

provisioned_link_count: Option<i32>

[Output Only] Number of links actually provisioned in this interconnect.

requested_link_count: Option<i32>

Target number of physical links in the link bundle, as requested by the customer.

satisfies_pzs: Option<bool>

[Output Only] Set to true if the resource satisfies the zone separation organization policy constraints and false otherwise. Defaults to false if the field is not present.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

state: Option<String>

[Output Only] The current state of Interconnect functionality, which can take one of the following values: - ACTIVE: The Interconnect is valid, turned up and ready to use. Attachments may be provisioned on this Interconnect. - UNPROVISIONED: The Interconnect has not completed turnup. No attachments may be provisioned on this Interconnect. - UNDER_MAINTENANCE: The Interconnect is undergoing internal maintenance. No attachments may be provisioned or updated on this Interconnect. Check the State enum for the list of possible values.

Implementations

impl Interconnect

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn customer_name(&self) -> &str

Returns the value of customer_name, or the default value if customer_name is unset.

pub fn noc_contact_email(&self) -> &str

Returns the value of noc_contact_email, or the default value if noc_contact_email is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn requested_link_count(&self) -> i32

Returns the value of requested_link_count, or the default value if requested_link_count is unset.

pub fn state(&self) -> &str

Returns the value of state, or the default value if state is unset.

pub fn operational_status(&self) -> &str

Returns the value of operational_status, or the default value if operational_status is unset.

pub fn peer_ip_address(&self) -> &str

Returns the value of peer_ip_address, or the default value if peer_ip_address is unset.

pub fn location(&self) -> &str

Returns the value of location, or the default value if location is unset.

pub fn provisioned_link_count(&self) -> i32

Returns the value of provisioned_link_count, or the default value if provisioned_link_count is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn google_ip_address(&self) -> &str

Returns the value of google_ip_address, or the default value if google_ip_address is unset.

pub fn admin_enabled(&self) -> bool

Returns the value of admin_enabled, or the default value if admin_enabled is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.

pub fn interconnect_type(&self) -> &str

Returns the value of interconnect_type, or the default value if interconnect_type is unset.

pub fn link_type(&self) -> &str

Returns the value of link_type, or the default value if link_type is unset.

pub fn google_reference_id(&self) -> &str

Returns the value of google_reference_id, or the default value if google_reference_id is unset.