Struct google_api_proto::google::cloud::compute::v1::InterconnectAttachment

pub struct InterconnectAttachment {

    pub admin_enabled: Option<bool>,
    pub bandwidth: Option<String>,
    pub candidate_ipv6_subnets: Vec<String>,
    pub candidate_subnets: Vec<String>,
    pub cloud_router_ip_address: Option<String>,
    pub cloud_router_ipv6_address: Option<String>,
    pub cloud_router_ipv6_interface_id: Option<String>,
    pub configuration_constraints: Option<InterconnectAttachmentConfigurationConstraints>,
    pub creation_timestamp: Option<String>,
    pub customer_router_ip_address: Option<String>,
    pub customer_router_ipv6_address: Option<String>,
    pub customer_router_ipv6_interface_id: Option<String>,
    pub dataplane_version: Option<i32>,
    pub description: Option<String>,
    pub edge_availability_domain: Option<String>,
    pub encryption: Option<String>,
    pub google_reference_id: Option<String>,
    pub id: Option<u64>,
    pub interconnect: Option<String>,
    pub ipsec_internal_addresses: Vec<String>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub mtu: Option<i32>,
    pub name: Option<String>,
    pub operational_status: Option<String>,
    pub pairing_key: Option<String>,
    pub partner_asn: Option<i64>,
    pub partner_metadata: Option<InterconnectAttachmentPartnerMetadata>,
    pub private_interconnect_info: Option<InterconnectAttachmentPrivateInfo>,
    pub region: Option<String>,
    pub remote_service: Option<String>,
    pub router: Option<String>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub stack_type: Option<String>,
    pub state: Option<String>,
    pub subnet_length: Option<i32>,
    pub type: Option<String>,
    pub vlan_tag8021q: Option<i32>,

}

Represents an Interconnect Attachment (VLAN) resource. You can use Interconnect attachments (VLANS) to connect your Virtual Private Cloud networks to your on-premises networks through an Interconnect. For more information, read Creating VLAN Attachments.

Fields
admin_enabled: Option<bool>

Determines whether this Attachment will carry packets. Not present for PARTNER_PROVIDER.

bandwidth: Option<String>

Provisioned bandwidth capacity for the interconnect attachment. For attachments of type DEDICATED, the user can set the bandwidth. For attachments of type PARTNER, the Google Partner that is operating the interconnect must set the bandwidth. Output only for PARTNER type, mutable for PARTNER_PROVIDER and DEDICATED, and can take one of the following values: - BPS_50M: 50 Mbit/s - BPS_100M: 100 Mbit/s - BPS_200M: 200 Mbit/s - BPS_300M: 300 Mbit/s - BPS_400M: 400 Mbit/s - BPS_500M: 500 Mbit/s - BPS_1G: 1 Gbit/s - BPS_2G: 2 Gbit/s - BPS_5G: 5 Gbit/s - BPS_10G: 10 Gbit/s - BPS_20G: 20 Gbit/s - BPS_50G: 50 Gbit/s Check the Bandwidth enum for the list of possible values.

candidate_ipv6_subnets: Vec<String>

This field is not available.

candidate_subnets: Vec<String>

Up to 16 candidate prefixes that can be used to restrict the allocation of cloudRouterIpAddress and customerRouterIpAddress for this attachment. All prefixes must be within link-local address space (169.254.0.0/16) and must be /29 or shorter (/28, /27, etc). Google will attempt to select an unused /29 from the supplied candidate prefix(es). The request will fail if all possible /29s are in use on Google’s edge. If not supplied, Google will randomly select an unused /29 from all of link-local space.

cloud_router_ip_address: Option<String>

[Output Only] IPv4 address + prefix length to be configured on Cloud Router Interface for this interconnect attachment.

cloud_router_ipv6_address: Option<String>

[Output Only] IPv6 address + prefix length to be configured on Cloud Router Interface for this interconnect attachment.

cloud_router_ipv6_interface_id: Option<String>

This field is not available.

configuration_constraints: Option<InterconnectAttachmentConfigurationConstraints>

[Output Only] Constraints for this attachment, if any. The attachment does not work if these constraints are not met.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

customer_router_ip_address: Option<String>

[Output Only] IPv4 address + prefix length to be configured on the customer router subinterface for this interconnect attachment.

customer_router_ipv6_address: Option<String>

[Output Only] IPv6 address + prefix length to be configured on the customer router subinterface for this interconnect attachment.

customer_router_ipv6_interface_id: Option<String>

This field is not available.

dataplane_version: Option<i32>

[Output Only] Dataplane version for this InterconnectAttachment. This field is only present for Dataplane version 2 and higher. Absence of this field in the API output indicates that the Dataplane is version 1.

description: Option<String>

An optional description of this resource.

edge_availability_domain: Option<String>

Desired availability domain for the attachment. Only available for type PARTNER, at creation time, and can take one of the following values: - AVAILABILITY_DOMAIN_ANY - AVAILABILITY_DOMAIN_1 - AVAILABILITY_DOMAIN_2 For improved reliability, customers should configure a pair of attachments, one per availability domain. The selected availability domain will be provided to the Partner via the pairing key, so that the provisioned circuit will lie in the specified domain. If not specified, the value will default to AVAILABILITY_DOMAIN_ANY. Check the EdgeAvailabilityDomain enum for the list of possible values.

encryption: Option<String>

Indicates the user-supplied encryption option of this VLAN attachment (interconnectAttachment). Can only be specified at attachment creation for PARTNER or DEDICATED attachments. Possible values are: - NONE - This is the default value, which means that the VLAN attachment carries unencrypted traffic. VMs are able to send traffic to, or receive traffic from, such a VLAN attachment. - IPSEC - The VLAN attachment carries only encrypted traffic that is encrypted by an IPsec device, such as an HA VPN gateway or third-party IPsec VPN. VMs cannot directly send traffic to, or receive traffic from, such a VLAN attachment. To use HA VPN over Cloud Interconnect, the VLAN attachment must be created with this option. Check the Encryption enum for the list of possible values.

google_reference_id: Option<String>

[Output Only] Google reference ID, to be used when raising support tickets with Google or otherwise to debug backend connectivity issues. [Deprecated] This field is not used.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

interconnect: Option<String>

URL of the underlying Interconnect object that this attachment’s traffic will traverse through.

ipsec_internal_addresses: Vec<String>

A list of URLs of addresses that have been reserved for the VLAN attachment. Used only for the VLAN attachment that has the encryption option as IPSEC. The addresses must be regional internal IP address ranges. When creating an HA VPN gateway over the VLAN attachment, if the attachment is configured to use a regional internal IP address, then the VPN gateway’s IP address is allocated from the IP address range specified here. For example, if the HA VPN gateway’s interface 0 is paired to this VLAN attachment, then a regional internal IP address for the VPN gateway interface 0 will be allocated from the IP address specified for this VLAN attachment. If this field is not specified when creating the VLAN attachment, then later on when creating an HA VPN gateway on this VLAN attachment, the HA VPN gateway’s IP address is allocated from the regional external IP address pool.

kind: Option<String>

[Output Only] Type of the resource. Always compute#interconnectAttachment for interconnect attachments.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this InterconnectAttachment, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve an InterconnectAttachment.

labels: BTreeMap<String, String>

Labels for this resource. These can only be added or modified by the setLabels method. Each label key/value pair must comply with RFC1035. Label values may be empty.

mtu: Option<i32>

Maximum Transmission Unit (MTU), in bytes, of packets passing through this interconnect attachment. Only 1440 and 1500 are allowed. If not specified, the value will default to 1440.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

operational_status: Option<String>

[Output Only] The current status of whether or not this interconnect attachment is functional, which can take one of the following values: - OS_ACTIVE: The attachment has been turned up and is ready to use. - OS_UNPROVISIONED: The attachment is not ready to use yet, because turnup is not complete. Check the OperationalStatus enum for the list of possible values.

pairing_key: Option<String>

[Output only for type PARTNER. Input only for PARTNER_PROVIDER. Not present for DEDICATED]. The opaque identifier of a PARTNER attachment used to initiate provisioning with a selected partner. Of the form “XXXXX/region/domain”

partner_asn: Option<i64>

Optional BGP ASN for the router supplied by a Layer 3 Partner if they configured BGP on behalf of the customer. Output only for PARTNER type, input only for PARTNER_PROVIDER, not available for DEDICATED.

partner_metadata: Option<InterconnectAttachmentPartnerMetadata>

Informational metadata about Partner attachments from Partners to display to customers. Output only for PARTNER type, mutable for PARTNER_PROVIDER, not available for DEDICATED.

private_interconnect_info: Option<InterconnectAttachmentPrivateInfo>

[Output Only] Information specific to an InterconnectAttachment. This property is populated if the interconnect that this is attached to is of type DEDICATED.

region: Option<String>

[Output Only] URL of the region where the regional interconnect attachment resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

remote_service: Option<String>

[Output Only] If the attachment is on a Cross-Cloud Interconnect connection, this field contains the interconnect’s remote location service provider. Example values: “Amazon Web Services” “Microsoft Azure”. The field is set only for attachments on Cross-Cloud Interconnect connections. Its value is copied from the InterconnectRemoteLocation remoteService field.

router: Option<String>

URL of the Cloud Router to be used for dynamic routing. This router must be in the same region as this InterconnectAttachment. The InterconnectAttachment will automatically connect the Interconnect to the network & region within which the Cloud Router is configured.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

stack_type: Option<String>

The stack type for this interconnect attachment to identify whether the IPv6 feature is enabled or not. If not specified, IPV4_ONLY will be used. This field can be both set at interconnect attachments creation and update interconnect attachment operations. Check the StackType enum for the list of possible values.

state: Option<String>

[Output Only] The current state of this attachment’s functionality. Enum values ACTIVE and UNPROVISIONED are shared by DEDICATED/PRIVATE, PARTNER, and PARTNER_PROVIDER interconnect attachments, while enum values PENDING_PARTNER, PARTNER_REQUEST_RECEIVED, and PENDING_CUSTOMER are used for only PARTNER and PARTNER_PROVIDER interconnect attachments. This state can take one of the following values: - ACTIVE: The attachment has been turned up and is ready to use. - UNPROVISIONED: The attachment is not ready to use yet, because turnup is not complete. - PENDING_PARTNER: A newly-created PARTNER attachment that has not yet been configured on the Partner side. - PARTNER_REQUEST_RECEIVED: A PARTNER attachment is in the process of provisioning after a PARTNER_PROVIDER attachment was created that references it. - PENDING_CUSTOMER: A PARTNER or PARTNER_PROVIDER attachment that is waiting for a customer to activate it. - DEFUNCT: The attachment was deleted externally and is no longer functional. This could be because the associated Interconnect was removed, or because the other side of a Partner attachment was deleted. Check the State enum for the list of possible values.

subnet_length: Option<i32>

Length of the IPv4 subnet mask. Allowed values: - 29 (default) - 30 The default value is 29, except for Cross-Cloud Interconnect connections that use an InterconnectRemoteLocation with a constraints.subnetLengthRange.min equal to 30. For example, connections that use an Azure remote location fall into this category. In these cases, the default value is 30, and requesting 29 returns an error. Where both 29 and 30 are allowed, 29 is preferred, because it gives Google Cloud Support more debugging visibility.

type: Option<String>

The type of interconnect attachment this is, which can take one of the following values: - DEDICATED: an attachment to a Dedicated Interconnect. - PARTNER: an attachment to a Partner Interconnect, created by the customer. - PARTNER_PROVIDER: an attachment to a Partner Interconnect, created by the partner. Check the Type enum for the list of possible values.

vlan_tag8021q: Option<i32>

The IEEE 802.1Q VLAN tag for this attachment, in the range 2-4093. Only specified at creation time.

Implementations

impl InterconnectAttachment

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn mtu(&self) -> i32

Returns the value of mtu, or the default value if mtu is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn type(&self) -> &str

Returns the value of type, or the default value if type is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn dataplane_version(&self) -> i32

Returns the value of dataplane_version, or the default value if dataplane_version is unset.

pub fn edge_availability_domain(&self) -> &str

Returns the value of edge_availability_domain, or the default value if edge_availability_domain is unset.

pub fn encryption(&self) -> &str

Returns the value of encryption, or the default value if encryption is unset.

pub fn state(&self) -> &str

Returns the value of state, or the default value if state is unset.

pub fn vlan_tag8021q(&self) -> i32

Returns the value of vlan_tag8021q, or the default value if vlan_tag8021q is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn router(&self) -> &str

Returns the value of router, or the default value if router is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn bandwidth(&self) -> &str

Returns the value of bandwidth, or the default value if bandwidth is unset.

pub fn operational_status(&self) -> &str

Returns the value of operational_status, or the default value if operational_status is unset.

pub fn interconnect(&self) -> &str

Returns the value of interconnect, or the default value if interconnect is unset.

pub fn subnet_length(&self) -> i32

Returns the value of subnet_length, or the default value if subnet_length is unset.

pub fn cloud_router_ip_address(&self) -> &str

Returns the value of cloud_router_ip_address, or the default value if cloud_router_ip_address is unset.

pub fn customer_router_ipv6_address(&self) -> &str

Returns the value of customer_router_ipv6_address, or the default value if customer_router_ipv6_address is unset.

pub fn customer_router_ip_address(&self) -> &str

Returns the value of customer_router_ip_address, or the default value if customer_router_ip_address is unset.

pub fn customer_router_ipv6_interface_id(&self) -> &str

Returns the value of customer_router_ipv6_interface_id, or the default value if customer_router_ipv6_interface_id is unset.

pub fn remote_service(&self) -> &str

Returns the value of remote_service, or the default value if remote_service is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn stack_type(&self) -> &str

Returns the value of stack_type, or the default value if stack_type is unset.

pub fn partner_asn(&self) -> i64

Returns the value of partner_asn, or the default value if partner_asn is unset.

pub fn pairing_key(&self) -> &str

Returns the value of pairing_key, or the default value if pairing_key is unset.

pub fn admin_enabled(&self) -> bool

Returns the value of admin_enabled, or the default value if admin_enabled is unset.

pub fn cloud_router_ipv6_address(&self) -> &str

Returns the value of cloud_router_ipv6_address, or the default value if cloud_router_ipv6_address is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.

pub fn cloud_router_ipv6_interface_id(&self) -> &str

Returns the value of cloud_router_ipv6_interface_id, or the default value if cloud_router_ipv6_interface_id is unset.

pub fn google_reference_id(&self) -> &str

Returns the value of google_reference_id, or the default value if google_reference_id is unset.