Struct google_api_proto::google::cloud::compute::v1::MachineImage

pub struct MachineImage {

    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub guest_flush: Option<bool>,
    pub id: Option<u64>,
    pub instance_properties: Option<InstanceProperties>,
    pub kind: Option<String>,
    pub machine_image_encryption_key: Option<CustomerEncryptionKey>,
    pub name: Option<String>,
    pub satisfies_pzi: Option<bool>,
    pub satisfies_pzs: Option<bool>,
    pub saved_disks: Vec<SavedDisk>,
    pub self_link: Option<String>,
    pub source_disk_encryption_keys: Vec<SourceDiskEncryptionKey>,
    pub source_instance: Option<String>,
    pub source_instance_properties: Option<SourceInstanceProperties>,
    pub status: Option<String>,
    pub storage_locations: Vec<String>,
    pub total_storage_bytes: Option<i64>,

}

Represents a machine image resource. A machine image is a Compute Engine resource that stores all the configuration, metadata, permissions, and data from one or more disks required to create a Virtual machine (VM) instance. For more information, see Machine images.

Fields
creation_timestamp: Option<String>

[Output Only] The creation timestamp for this machine image in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

guest_flush: Option<bool>

[Input Only] Whether to attempt an application consistent machine image by informing the OS to prepare for the snapshot process.

id: Option<u64>

[Output Only] A unique identifier for this machine image. The server defines this identifier.

instance_properties: Option<InstanceProperties>

[Output Only] Properties of source instance

kind: Option<String>

[Output Only] The resource type, which is always compute#machineImage for machine image.

machine_image_encryption_key: Option<CustomerEncryptionKey>

Encrypts the machine image using a customer-supplied encryption key. After you encrypt a machine image using a customer-supplied key, you must provide the same key if you use the machine image later. For example, you must provide the encryption key when you create an instance from the encrypted machine image in a future request. Customer-supplied encryption keys do not protect access to metadata of the machine image. If you do not provide an encryption key when creating the machine image, then the machine image will be encrypted using an automatically generated key and you do not need to provide a key to use the machine image later.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

satisfies_pzi: Option<bool>

Output only. Reserved for future use.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

saved_disks: Vec<SavedDisk>

An array of Machine Image specific properties for disks attached to the source instance

self_link: Option<String>

[Output Only] The URL for this machine image. The server defines this URL.

source_disk_encryption_keys: Vec<SourceDiskEncryptionKey>

[Input Only] The customer-supplied encryption key of the disks attached to the source instance. Required if the source disk is protected by a customer-supplied encryption key.

source_instance: Option<String>

The source instance used to create the machine image. You can provide this as a partial or full URL to the resource. For example, the following are valid values: - https://www.googleapis.com/compute/v1/projects/project/zones/zone /instances/instance - projects/project/zones/zone/instances/instance

source_instance_properties: Option<SourceInstanceProperties>

[Output Only] DEPRECATED: Please use instance_properties instead for source instance related properties. New properties will not be added to this field.

status: Option<String>

[Output Only] The status of the machine image. One of the following values: INVALID, CREATING, READY, DELETING, and UPLOADING. Check the Status enum for the list of possible values.

storage_locations: Vec<String>

The regional or multi-regional Cloud Storage bucket location where the machine image is stored.

total_storage_bytes: Option<i64>

[Output Only] Total size of the storage used by the machine image.

Implementations

impl MachineImage

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn total_storage_bytes(&self) -> i64

Returns the value of total_storage_bytes, or the default value if total_storage_bytes is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn guest_flush(&self) -> bool

Returns the value of guest_flush, or the default value if guest_flush is unset.

pub fn source_instance(&self) -> &str

Returns the value of source_instance, or the default value if source_instance is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn satisfies_pzi(&self) -> bool

Returns the value of satisfies_pzi, or the default value if satisfies_pzi is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.