Struct google_api_proto::google::cloud::compute::v1::Snapshot

pub struct Snapshot {

    pub architecture: Option<String>,
    pub auto_created: Option<bool>,
    pub chain_name: Option<String>,
    pub creation_size_bytes: Option<i64>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub disk_size_gb: Option<i64>,
    pub download_bytes: Option<i64>,
    pub guest_os_features: Vec<GuestOsFeature>,
    pub id: Option<u64>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: BTreeMap<String, String>,
    pub license_codes: Vec<i64>,
    pub licenses: Vec<String>,
    pub location_hint: Option<String>,
    pub name: Option<String>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub snapshot_encryption_key: Option<CustomerEncryptionKey>,
    pub snapshot_type: Option<String>,
    pub source_disk: Option<String>,
    pub source_disk_encryption_key: Option<CustomerEncryptionKey>,
    pub source_disk_for_recovery_checkpoint: Option<String>,
    pub source_disk_id: Option<String>,
    pub source_snapshot_schedule_policy: Option<String>,
    pub source_snapshot_schedule_policy_id: Option<String>,
    pub status: Option<String>,
    pub storage_bytes: Option<i64>,
    pub storage_bytes_status: Option<String>,
    pub storage_locations: Vec<String>,

}

Represents a Persistent Disk Snapshot resource. You can use snapshots to back up data on a regular interval. For more information, read Creating persistent disk snapshots.

Fields
architecture: Option<String>

[Output Only] The architecture of the snapshot. Valid values are ARM64 or X86_64. Check the Architecture enum for the list of possible values.

auto_created: Option<bool>

[Output Only] Set to true if snapshots are automatically created by applying resource policy on the target disk.

chain_name: Option<String>

Creates the new snapshot in the snapshot chain labeled with the specified name. The chain name must be 1-63 characters long and comply with RFC1035. This is an uncommon option only for advanced service owners who needs to create separate snapshot chains, for example, for chargeback tracking. When you describe your snapshot resource, this field is visible only if it has a non-empty value.

creation_size_bytes: Option<i64>

[Output Only] Size in bytes of the snapshot at creation time.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disk_size_gb: Option<i64>

[Output Only] Size of the source disk, specified in GB.

download_bytes: Option<i64>

[Output Only] Number of bytes downloaded to restore a snapshot to a disk.

guest_os_features: Vec<GuestOsFeature>

[Output Only] A list of features to enable on the guest operating system. Applicable only for bootable images. Read Enabling guest operating system features to see a list of available options.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#snapshot for Snapshot resources.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this snapshot, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a snapshot.

labels: BTreeMap<String, String>

Labels to apply to this snapshot. These can be later modified by the setLabels method. Label values may be empty.

license_codes: Vec<i64>

[Output Only] Integer license codes indicating which licenses are attached to this snapshot.

licenses: Vec<String>

[Output Only] A list of public visible licenses that apply to this snapshot. This can be because the original image had licenses attached (such as a Windows image).

location_hint: Option<String>

An opaque location hint used to place the snapshot close to other resources. This field is for use by internal tools that use the public API.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

snapshot_encryption_key: Option<CustomerEncryptionKey>

Encrypts the snapshot using a customer-supplied encryption key. After you encrypt a snapshot using a customer-supplied key, you must provide the same key if you use the snapshot later. For example, you must provide the encryption key when you create a disk from the encrypted snapshot in a future request. Customer-supplied encryption keys do not protect access to metadata of the snapshot. If you do not provide an encryption key when creating the snapshot, then the snapshot will be encrypted using an automatically generated key and you do not need to provide a key to use the snapshot later.

snapshot_type: Option<String>

Indicates the type of the snapshot. Check the SnapshotType enum for the list of possible values.

source_disk: Option<String>

The source disk used to create this snapshot.

source_disk_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source disk. Required if the source disk is protected by a customer-supplied encryption key.

source_disk_for_recovery_checkpoint: Option<String>

The source disk whose recovery checkpoint will be used to create this snapshot.

source_disk_id: Option<String>

[Output Only] The ID value of the disk used to create this snapshot. This value may be used to determine whether the snapshot was taken from the current or a previous instance of a given disk name.

source_snapshot_schedule_policy: Option<String>

[Output Only] URL of the resource policy which created this scheduled snapshot.

source_snapshot_schedule_policy_id: Option<String>

[Output Only] ID of the resource policy which created this scheduled snapshot.

status: Option<String>

[Output Only] The status of the snapshot. This can be CREATING, DELETING, FAILED, READY, or UPLOADING. Check the Status enum for the list of possible values.

storage_bytes: Option<i64>

[Output Only] A size of the storage used by the snapshot. As snapshots share storage, this number is expected to change with snapshot creation/deletion.

storage_bytes_status: Option<String>

[Output Only] An indicator whether storageBytes is in a stable state or it is being adjusted as a result of shared storage reallocation. This status can either be UPDATING, meaning the size of the snapshot is being updated, or UP_TO_DATE, meaning the size of the snapshot is up-to-date. Check the StorageBytesStatus enum for the list of possible values.

storage_locations: Vec<String>

Cloud Storage bucket storage location of the snapshot (regional or multi-regional).

Implementations

impl Snapshot

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn chain_name(&self) -> &str

Returns the value of chain_name, or the default value if chain_name is unset.

pub fn source_snapshot_schedule_policy_id(&self) -> &str

Returns the value of source_snapshot_schedule_policy_id, or the default value if source_snapshot_schedule_policy_id is unset.

pub fn snapshot_type(&self) -> &str

Returns the value of snapshot_type, or the default value if snapshot_type is unset.

pub fn creation_size_bytes(&self) -> i64

Returns the value of creation_size_bytes, or the default value if creation_size_bytes is unset.

pub fn label_fingerprint(&self) -> &str

Returns the value of label_fingerprint, or the default value if label_fingerprint is unset.

pub fn status(&self) -> &str

Returns the value of status, or the default value if status is unset.

pub fn source_snapshot_schedule_policy(&self) -> &str

Returns the value of source_snapshot_schedule_policy, or the default value if source_snapshot_schedule_policy is unset.

pub fn architecture(&self) -> &str

Returns the value of architecture, or the default value if architecture is unset.

pub fn disk_size_gb(&self) -> i64

Returns the value of disk_size_gb, or the default value if disk_size_gb is unset.

pub fn location_hint(&self) -> &str

Returns the value of location_hint, or the default value if location_hint is unset.

pub fn source_disk_for_recovery_checkpoint(&self) -> &str

Returns the value of source_disk_for_recovery_checkpoint, or the default value if source_disk_for_recovery_checkpoint is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn storage_bytes(&self) -> i64

Returns the value of storage_bytes, or the default value if storage_bytes is unset.

pub fn download_bytes(&self) -> i64

Returns the value of download_bytes, or the default value if download_bytes is unset.

pub fn source_disk(&self) -> &str

Returns the value of source_disk, or the default value if source_disk is unset.

pub fn source_disk_id(&self) -> &str

Returns the value of source_disk_id, or the default value if source_disk_id is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn auto_created(&self) -> bool

Returns the value of auto_created, or the default value if auto_created is unset.

pub fn satisfies_pzs(&self) -> bool

Returns the value of satisfies_pzs, or the default value if satisfies_pzs is unset.

pub fn storage_bytes_status(&self) -> &str

Returns the value of storage_bytes_status, or the default value if storage_bytes_status is unset.