Struct google_api_proto::google::cloud::compute::v1::Subnetwork

pub struct Subnetwork {

    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub enable_flow_logs: Option<bool>,
    pub external_ipv6_prefix: Option<String>,
    pub fingerprint: Option<String>,
    pub gateway_address: Option<String>,
    pub id: Option<u64>,
    pub internal_ipv6_prefix: Option<String>,
    pub ip_cidr_range: Option<String>,
    pub ipv6_access_type: Option<String>,
    pub ipv6_cidr_range: Option<String>,
    pub kind: Option<String>,
    pub log_config: Option<SubnetworkLogConfig>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub private_ip_google_access: Option<bool>,
    pub private_ipv6_google_access: Option<String>,
    pub purpose: Option<String>,
    pub region: Option<String>,
    pub role: Option<String>,
    pub secondary_ip_ranges: Vec<SubnetworkSecondaryRange>,
    pub self_link: Option<String>,
    pub stack_type: Option<String>,
    pub state: Option<String>,

}

Represents a Subnetwork resource. A subnetwork (also known as a subnet) is a logical partition of a Virtual Private Cloud network with one primary IP range and zero or more secondary IP ranges. For more information, read Virtual Private Cloud (VPC) Network.

Fields
creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource. This field can be set only at resource creation time.

enable_flow_logs: Option<bool>

Whether to enable flow logging for this subnetwork. If this field is not explicitly set, it will not appear in get listings. If not set the default behavior is determined by the org policy, if there is no org policy specified, then it will default to disabled. This field isn’t supported if the subnet purpose field is set to REGIONAL_MANAGED_PROXY.

external_ipv6_prefix: Option<String>

The external IPv6 address range that is owned by this subnetwork.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a Subnetwork. An up-to-date fingerprint must be provided in order to update the Subnetwork, otherwise the request will fail with error 412 conditionNotMet. To see the latest fingerprint, make a get() request to retrieve a Subnetwork.

gateway_address: Option<String>

[Output Only] The gateway address for default routes to reach destination addresses outside this subnetwork.

id: Option<u64>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

internal_ipv6_prefix: Option<String>

[Output Only] The internal IPv6 address range that is assigned to this subnetwork.

ip_cidr_range: Option<String>

The range of internal addresses that are owned by this subnetwork. Provide this property when you create the subnetwork. For example, 10.0.0.0/8 or 100.64.0.0/10. Ranges must be unique and non-overlapping within a network. Only IPv4 is supported. This field is set at resource creation time. The range can be any range listed in the Valid ranges list. The range can be expanded after creation using expandIpCidrRange.

ipv6_access_type: Option<String>

The access type of IPv6 address this subnet holds. It’s immutable and can only be specified during creation or the first time the subnet is updated into IPV4_IPV6 dual stack. Check the Ipv6AccessType enum for the list of possible values.

ipv6_cidr_range: Option<String>

[Output Only] This field is for internal use.

kind: Option<String>

[Output Only] Type of the resource. Always compute#subnetwork for Subnetwork resources.

log_config: Option<SubnetworkLogConfig>

This field denotes the VPC flow logging options for this subnetwork. If logging is enabled, logs are exported to Cloud Logging.

name: Option<String>

The name of the resource, provided by the client when initially creating the resource. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z](\[-a-z0-9\]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network: Option<String>

The URL of the network to which this subnetwork belongs, provided by the client when initially creating the subnetwork. This field can be set only at resource creation time.

private_ip_google_access: Option<bool>

Whether the VMs in this subnet can access Google services without assigned external IP addresses. This field can be both set at resource creation time and updated using setPrivateIpGoogleAccess.

private_ipv6_google_access: Option<String>

This field is for internal use. This field can be both set at resource creation time and updated using patch. Check the PrivateIpv6GoogleAccess enum for the list of possible values.

purpose: Option<String>

The purpose of the resource. This field can be either PRIVATE, REGIONAL_MANAGED_PROXY, PRIVATE_SERVICE_CONNECT, or INTERNAL_HTTPS_LOAD_BALANCER. PRIVATE is the default purpose for user-created subnets or subnets that are automatically created in auto mode networks. A subnet with purpose set to REGIONAL_MANAGED_PROXY is a user-created subnetwork that is reserved for regional Envoy-based load balancers. A subnet with purpose set to PRIVATE_SERVICE_CONNECT is used to publish services using Private Service Connect. A subnet with purpose set to INTERNAL_HTTPS_LOAD_BALANCER is a proxy-only subnet that can be used only by regional internal HTTP(S) load balancers. Note that REGIONAL_MANAGED_PROXY is the preferred setting for all regional Envoy load balancers. If unspecified, the subnet purpose defaults to PRIVATE. The enableFlowLogs field isn’t supported if the subnet purpose field is set to REGIONAL_MANAGED_PROXY. Check the Purpose enum for the list of possible values.

region: Option<String>

URL of the region where the Subnetwork resides. This field can be set only at resource creation time.

role: Option<String>

The role of subnetwork. Currently, this field is only used when purpose = REGIONAL_MANAGED_PROXY. The value can be set to ACTIVE or BACKUP. An ACTIVE subnetwork is one that is currently being used for Envoy-based load balancers in a region. A BACKUP subnetwork is one that is ready to be promoted to ACTIVE or is currently draining. This field can be updated with a patch request. Check the Role enum for the list of possible values.

secondary_ip_ranges: Vec<SubnetworkSecondaryRange>

An array of configurations for secondary IP ranges for VM instances contained in this subnetwork. The primary IP of such VM must belong to the primary ipCidrRange of the subnetwork. The alias IPs may belong to either primary or secondary ranges. This field can be updated with a patch request.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

stack_type: Option<String>

The stack type for the subnet. If set to IPV4_ONLY, new VMs in the subnet are assigned IPv4 addresses only. If set to IPV4_IPV6, new VMs in the subnet can be assigned both IPv4 and IPv6 addresses. If not specified, IPV4_ONLY is used. This field can be both set at resource creation time and updated using patch. Check the StackType enum for the list of possible values.

state: Option<String>

[Output Only] The state of the subnetwork, which can be one of the following values: READY: Subnetwork is created and ready to use DRAINING: only applicable to subnetworks that have the purpose set to INTERNAL_HTTPS_LOAD_BALANCER and indicates that connections to the load balancer are being drained. A subnetwork that is draining cannot be used or modified until it reaches a status of READY Check the State enum for the list of possible values.

Implementations

impl Subnetwork

pub fn id(&self) -> u64

Returns the value of id, or the default value if id is unset.

pub fn kind(&self) -> &str

Returns the value of kind, or the default value if kind is unset.

pub fn name(&self) -> &str

Returns the value of name, or the default value if name is unset.

pub fn role(&self) -> &str

Returns the value of role, or the default value if role is unset.

pub fn creation_timestamp(&self) -> &str

Returns the value of creation_timestamp, or the default value if creation_timestamp is unset.

pub fn private_ipv6_google_access(&self) -> &str

Returns the value of private_ipv6_google_access, or the default value if private_ipv6_google_access is unset.

pub fn ip_cidr_range(&self) -> &str

Returns the value of ip_cidr_range, or the default value if ip_cidr_range is unset.

pub fn state(&self) -> &str

Returns the value of state, or the default value if state is unset.

pub fn region(&self) -> &str

Returns the value of region, or the default value if region is unset.

pub fn external_ipv6_prefix(&self) -> &str

Returns the value of external_ipv6_prefix, or the default value if external_ipv6_prefix is unset.

pub fn enable_flow_logs(&self) -> bool

Returns the value of enable_flow_logs, or the default value if enable_flow_logs is unset.

pub fn network(&self) -> &str

Returns the value of network, or the default value if network is unset.

pub fn fingerprint(&self) -> &str

Returns the value of fingerprint, or the default value if fingerprint is unset.

pub fn ipv6_cidr_range(&self) -> &str

Returns the value of ipv6_cidr_range, or the default value if ipv6_cidr_range is unset.

pub fn purpose(&self) -> &str

Returns the value of purpose, or the default value if purpose is unset.

pub fn private_ip_google_access(&self) -> bool

Returns the value of private_ip_google_access, or the default value if private_ip_google_access is unset.

pub fn description(&self) -> &str

Returns the value of description, or the default value if description is unset.

pub fn stack_type(&self) -> &str

Returns the value of stack_type, or the default value if stack_type is unset.

pub fn self_link(&self) -> &str

Returns the value of self_link, or the default value if self_link is unset.

pub fn gateway_address(&self) -> &str

Returns the value of gateway_address, or the default value if gateway_address is unset.

pub fn ipv6_access_type(&self) -> &str

Returns the value of ipv6_access_type, or the default value if ipv6_access_type is unset.

pub fn internal_ipv6_prefix(&self) -> &str

Returns the value of internal_ipv6_prefix, or the default value if internal_ipv6_prefix is unset.