Module google_api_proto::google::cloud::config::v1

Modules

config_client

Generated client implementations.

delete_deployment_request

Nested message and enum types in DeleteDeploymentRequest.

deployment

Nested message and enum types in Deployment.

deployment_operation_metadata

Nested message and enum types in DeploymentOperationMetadata.

operation_metadata

Nested message and enum types in OperationMetadata.

resource

Nested message and enum types in Resource.

revision

Nested message and enum types in Revision.

terraform_blueprint

Nested message and enum types in TerraformBlueprint.

Structs

ApplyResults

Outputs and artifacts from applying a deployment.

CreateDeploymentRequest

DeleteDeploymentRequest

DeleteStatefileRequest

A request to delete a state file passed to a ‘DeleteStatefile’ call.

Deployment

A Deployment is a group of resources and configs managed and provisioned by Infra Manager.

DeploymentOperationMetadata

Ephemeral metadata content describing the state of a deployment operation.

ExportDeploymentStatefileRequest

A request to export a state file passed to a ‘ExportDeploymentStatefile’ call.

ExportLockInfoRequest

A request to get a state file lock info passed to a ‘ExportLockInfo’ call.

ExportRevisionStatefileRequest

A request to export a state file passed to a ‘ExportRevisionStatefile’ call.

GetDeploymentRequest

GetResourceRequest

A request to get a Resource from a ‘GetResource’ call.

GetRevisionRequest

A request to get a Revision from a ‘GetRevision’ call.

GitSource

A set of files in a Git repository.

ImportStatefileRequest

A request to import a state file passed to a ‘ImportStatefile’ call.

ListDeploymentsRequest

ListDeploymentsResponse

ListResourcesRequest

A request to list Resources passed to a ‘ListResources’ call.

ListResourcesResponse

A response to a ‘ListResources’ call. Contains a list of Resources.

ListRevisionsRequest

A request to list Revisions passed to a ‘ListRevisions’ call.

ListRevisionsResponse

A response to a ‘ListRevisions’ call. Contains a list of Revisions.

LockDeploymentRequest

A request to lock a deployment passed to a ‘LockDeployment’ call.

LockInfo

Details about the lock which locked the deployment.

OperationMetadata

Represents the metadata of the long-running operation.

Resource

Resource represents a Google Cloud Platform resource actuated by IM. Resources are child resources of Revisions.

ResourceCaiInfo

CAI info of a Resource.

ResourceTerraformInfo

Terraform info of a Resource.

Revision

A child resource of a Deployment generated by a ‘CreateDeployment’ or ‘UpdateDeployment’ call. Each Revision contains metadata pertaining to a snapshot of a particular Deployment.

Statefile

Contains info about a Terraform state file

TerraformBlueprint

TerraformBlueprint describes the source of a Terraform root module which describes the resources and configs to be deployed.

TerraformError

Errors encountered during actuation using Terraform

TerraformOutput

Describes a Terraform output.

TerraformVariable

A Terraform input variable.

UnlockDeploymentRequest

A request to unlock a state file passed to a ‘UnlockDeployment’ call.

UpdateDeploymentRequest