Module google_api_proto::google::cloud::config::v1

Modules

config_client

Generated client implementations.

delete_deployment_request

Nested message and enum types in DeleteDeploymentRequest.

deployment

Nested message and enum types in Deployment.

deployment_operation_metadata

Nested message and enum types in DeploymentOperationMetadata.

operation_metadata

Nested message and enum types in OperationMetadata.

preview

Nested message and enum types in Preview.

preview_operation_metadata

Nested message and enum types in PreviewOperationMetadata.

resource

Nested message and enum types in Resource.

revision

Nested message and enum types in Revision.

terraform_blueprint

Nested message and enum types in TerraformBlueprint.

terraform_version

Nested message and enum types in TerraformVersion.

Structs

ApplyResults

Outputs and artifacts from applying a deployment.

CreateDeploymentRequest

CreatePreviewRequest

A request to create a preview.

DeleteDeploymentRequest

DeletePreviewRequest

A request to delete a preview.

DeleteStatefileRequest

A request to delete a state file passed to a ‘DeleteStatefile’ call.

Deployment

A Deployment is a group of resources and configs managed and provisioned by Infra Manager.

DeploymentOperationMetadata

Ephemeral metadata content describing the state of a deployment operation.

ExportDeploymentStatefileRequest

A request to export a state file passed to a ‘ExportDeploymentStatefile’ call.

ExportLockInfoRequest

A request to get a state file lock info passed to a ‘ExportLockInfo’ call.

ExportPreviewResultRequest

A request to export preview results.

ExportPreviewResultResponse

A response to ExportPreviewResult call. Contains preview results.

ExportRevisionStatefileRequest

A request to export a state file passed to a ‘ExportRevisionStatefile’ call.

GetDeploymentRequest

GetPreviewRequest

A request to get details about a preview.

GetResourceRequest

A request to get a Resource from a ‘GetResource’ call.

GetRevisionRequest

A request to get a Revision from a ‘GetRevision’ call.

GetTerraformVersionRequest

The request message for the GetTerraformVersion method.

GitSource

A set of files in a Git repository.

ImportStatefileRequest

A request to import a state file passed to a ‘ImportStatefile’ call.

ListDeploymentsRequest

ListDeploymentsResponse

ListPreviewsRequest

A request to list all previews for a given project and location.

ListPreviewsResponse

A response to a ListPreviews call. Contains a list of Previews.

ListResourcesRequest

A request to list Resources passed to a ‘ListResources’ call.

ListResourcesResponse

A response to a ‘ListResources’ call. Contains a list of Resources.

ListRevisionsRequest

A request to list Revisions passed to a ‘ListRevisions’ call.

ListRevisionsResponse

A response to a ‘ListRevisions’ call. Contains a list of Revisions.

ListTerraformVersionsRequest

The request message for the ListTerraformVersions method.

ListTerraformVersionsResponse

The response message for the ListTerraformVersions method.

LockDeploymentRequest

A request to lock a deployment passed to a ‘LockDeployment’ call.

LockInfo

Details about the lock which locked the deployment.

OperationMetadata

Represents the metadata of the long-running operation.

Preview

A preview represents a set of actions Infra Manager would perform to move the resources towards the desired state as specified in the configuration.

PreviewArtifacts

Artifacts created by preview.

PreviewOperationMetadata

Ephemeral metadata content describing the state of a preview operation.

PreviewResult

Contains a signed Cloud Storage URLs.

Resource

Resource represents a Google Cloud Platform resource actuated by IM. Resources are child resources of Revisions.

ResourceCaiInfo

CAI info of a Resource.

ResourceTerraformInfo

Terraform info of a Resource.

Revision

A child resource of a Deployment generated by a ‘CreateDeployment’ or ‘UpdateDeployment’ call. Each Revision contains metadata pertaining to a snapshot of a particular Deployment.

Statefile

Contains info about a Terraform state file

TerraformBlueprint

TerraformBlueprint describes the source of a Terraform root module which describes the resources and configs to be deployed.

TerraformError

Errors encountered during actuation using Terraform

TerraformOutput

Describes a Terraform output.

TerraformVariable

A Terraform input variable.

TerraformVersion

A TerraformVersion represents the support state the corresponding Terraform version.

UnlockDeploymentRequest

A request to unlock a state file passed to a ‘UnlockDeployment’ call.

UpdateDeploymentRequest

Enums

QuotaValidation

Enum values to control quota checks for resources in terraform configuration files.