Module google_api_proto::google::cloud::contactcenterinsights::v1

Modules

analysis_result

Nested message and enum types in AnalysisResult.

annotation_boundary

Nested message and enum types in AnnotationBoundary.

annotator_selector

Nested message and enum types in AnnotatorSelector.

answer_feedback

Nested message and enum types in AnswerFeedback.

calculate_stats_response

Nested message and enum types in CalculateStatsResponse.

call_annotation

Nested message and enum types in CallAnnotation.

contact_center_insights_client

Generated client implementations.

conversation

Nested message and enum types in Conversation.

conversation_data_source

Nested message and enum types in ConversationDataSource.

conversation_participant

Nested message and enum types in ConversationParticipant.

entity

Nested message and enum types in Entity.

entity_mention_data

Nested message and enum types in EntityMentionData.

export_insights_data_request

Nested message and enum types in ExportInsightsDataRequest.

ingest_conversations_metadata

Nested message and enum types in IngestConversationsMetadata.

ingest_conversations_request

Nested message and enum types in IngestConversationsRequest.

issue_model

Nested message and enum types in IssueModel.

issue_model_label_stats

Nested message and enum types in IssueModelLabelStats.

phrase_match_rule_config

Nested message and enum types in PhraseMatchRuleConfig.

phrase_match_rule_group

Nested message and enum types in PhraseMatchRuleGroup.

phrase_matcher

Nested message and enum types in PhraseMatcher.

runtime_annotation

Nested message and enum types in RuntimeAnnotation.

settings

Nested message and enum types in Settings.

Structs

Analysis

The analysis resource.

AnalysisResult

The result of an analysis.

AnnotationBoundary

A point in a conversation that marks the start or the end of an annotation.

AnnotatorSelector

Selector of all available annotators and phrase matchers to run.

AnswerFeedback

The feedback that the customer has about a certain answer in the conversation.

ArticleSuggestionData

Agent Assist Article Suggestion data.

BulkAnalyzeConversationsMetadata

The metadata for a bulk analyze conversations operation.

BulkAnalyzeConversationsRequest

The request to analyze conversations in bulk.

BulkAnalyzeConversationsResponse

The response for a bulk analyze conversations operation.

BulkDeleteConversationsMetadata

The metadata for a bulk delete conversations operation.

BulkDeleteConversationsRequest

The request to delete conversations in bulk.

BulkDeleteConversationsResponse

The response for a bulk delete conversations operation.

CalculateIssueModelStatsRequest

Request to get statistics of an issue model.

CalculateIssueModelStatsResponse

Response of querying an issue model’s statistics.

CalculateStatsRequest

The request for calculating conversation statistics.

CalculateStatsResponse

The response for calculating conversation statistics.

CallAnnotation

A piece of metadata that applies to a window of a call.

Conversation

The conversation resource.

ConversationDataSource

The conversation source, which is a combination of transcript and audio.

ConversationLevelSentiment

One channel of conversation-level sentiment data.

ConversationParticipant

The call participant speaking for a given utterance.

ConversationSummarizationSuggestionData

Conversation summarization suggestion data.

CreateAnalysisOperationMetadata

Metadata for a create analysis operation.

CreateAnalysisRequest

The request to create an analysis.

CreateConversationRequest

Request to create a conversation.

CreateIssueModelMetadata

Metadata for creating an issue model.

CreateIssueModelRequest

The request to create an issue model.

CreatePhraseMatcherRequest

Request to create a phrase matcher.

CreateViewRequest

The request to create a view.

DeleteAnalysisRequest

The request to delete an analysis.

DeleteConversationRequest

The request to delete a conversation.

DeleteIssueModelMetadata

Metadata for deleting an issue model.

DeleteIssueModelRequest

The request to delete an issue model.

DeleteIssueRequest

The request to delete an issue.

DeletePhraseMatcherRequest

The request to delete a phrase matcher.

DeleteViewRequest

The request to delete a view.

DeployIssueModelMetadata

Metadata for deploying an issue model.

DeployIssueModelRequest

The request to deploy an issue model.

DeployIssueModelResponse

The response to deploy an issue model.

DialogflowIntent

The data for a Dialogflow intent. Represents a detected intent in the conversation, e.g. MAKES_PROMISE.

DialogflowInteractionData

Dialogflow interaction data.

DialogflowSource

A Dialogflow source of conversation data.

Entity

The data for an entity annotation. Represents a phrase in the conversation that is a known entity, such as a person, an organization, or location.

EntityMentionData

The data for an entity mention annotation. This represents a mention of an Entity in the conversation.

ExactMatchConfig

Exact match configuration.

ExportInsightsDataMetadata

Metadata for an export insights operation.

ExportInsightsDataRequest

The request to export insights.

ExportInsightsDataResponse

Response for an export insights operation.

FaqAnswerData

Agent Assist frequently-asked-question answer data.

GcsSource

A Cloud Storage source of conversation data.

GetAnalysisRequest

The request to get an analysis.

GetConversationRequest

The request to get a conversation.

GetIssueModelRequest

The request to get an issue model.

GetIssueRequest

The request to get an issue.

GetPhraseMatcherRequest

The request to get a a phrase matcher.

GetSettingsRequest

The request to get project-level settings.

GetViewRequest

The request to get a view.

HoldData

The data for a hold annotation.

IngestConversationsMetadata

The metadata for an IngestConversations operation.

IngestConversationsRequest

The request to ingest conversations.

IngestConversationsResponse

The response to an IngestConversations operation.

Intent

The data for an intent. Represents a detected intent in the conversation, for example MAKES_PROMISE.

IntentMatchData

The data for an intent match. Represents an intent match for a text segment in the conversation. A text segment can be part of a sentence, a complete sentence, or an utterance with multiple sentences.

InterruptionData

The data for an interruption annotation.

Issue

The issue resource.

IssueAssignment

Information about the issue.

IssueMatchData

The data for an issue match annotation.

IssueModel

The issue model resource.

IssueModelLabelStats

Aggregated statistics about an issue model.

IssueModelResult

Issue Modeling result on a conversation.

ListAnalysesRequest

The request to list analyses.

ListAnalysesResponse

The response to list analyses.

ListConversationsRequest

Request to list conversations.

ListConversationsResponse

The response of listing conversations.

ListIssueModelsRequest

Request to list issue models.

ListIssueModelsResponse

The response of listing issue models.

ListIssuesRequest

Request to list issues.

ListIssuesResponse

The response of listing issues.

ListPhraseMatchersRequest

Request to list phrase matchers.

ListPhraseMatchersResponse

The response of listing phrase matchers.

ListViewsRequest

The request to list views.

ListViewsResponse

The response of listing views.

PhraseMatchData

The data for a matched phrase matcher. Represents information identifying a phrase matcher for a given match.

PhraseMatchRule

The data for a phrase match rule.

PhraseMatchRuleConfig

Configuration information of a phrase match rule.

PhraseMatchRuleGroup

A message representing a rule in the phrase matcher.

PhraseMatcher

The phrase matcher resource.

RedactionConfig

DLP resources used for redaction while ingesting conversations.

RuntimeAnnotation

An annotation that was generated during the customer and agent interaction.

SentimentData

The data for a sentiment annotation.

Settings

The settings resource.

SilenceData

The data for a silence annotation.

SmartComposeSuggestionData

Agent Assist Smart Compose suggestion data.

SmartReplyData

Agent Assist Smart Reply data.

SpeechConfig

Speech-to-Text configuration.

UndeployIssueModelMetadata

Metadata for undeploying an issue model.

UndeployIssueModelRequest

The request to undeploy an issue model.

UndeployIssueModelResponse

The response to undeploy an issue model.

UpdateConversationRequest

The request to update a conversation.

UpdateIssueModelRequest

The request to update an issue model.

UpdateIssueRequest

The request to update an issue.

UpdatePhraseMatcherRequest

The request to update a phrase matcher.

UpdateSettingsRequest

The request to update project-level settings.

UpdateViewRequest

The request to update a view.

UploadConversationMetadata

The metadata for an UploadConversation operation.

UploadConversationRequest

Request to upload a conversation.

View

The View resource.

Enums

ConversationView

Represents the options for viewing a conversation.