Module google_api_proto::google::cloud::contentwarehouse::v1

Modules

access_control_action

Nested message and enum types in AccessControlAction.

action

Nested message and enum types in Action.

action_output

Nested message and enum types in ActionOutput.

document

Nested message and enum types in Document.

document_link

Nested message and enum types in DocumentLink.

document_link_service_client

Generated client implementations.

document_schema_service_client

Generated client implementations.

document_service_client

Generated client implementations.

file_type_filter

Nested message and enum types in FileTypeFilter.

histogram_query_property_name_filter

Nested message and enum types in HistogramQueryPropertyNameFilter.

pipeline_service_client

Generated client implementations.

property

Nested message and enum types in Property.

property_definition

Nested message and enum types in PropertyDefinition.

qa_result

Nested message and enum types in QAResult.

rule

Nested message and enum types in Rule.

rule_set_service_client

Generated client implementations.

run_pipeline_metadata

Nested message and enum types in RunPipelineMetadata.

run_pipeline_request

Nested message and enum types in RunPipelineRequest.

search_documents_request

Nested message and enum types in SearchDocumentsRequest.

search_documents_response

Nested message and enum types in SearchDocumentsResponse.

synonym_set

Nested message and enum types in SynonymSet.

synonym_set_service_client

Generated client implementations.

time_filter

Nested message and enum types in TimeFilter.

timestamp_value

Nested message and enum types in TimestampValue.

value

Nested message and enum types in Value.

Structs

AccessControlAction

Represents the action responsible for access control list management operations.

Action

Represents the action triggered by Rule Engine when the rule is true.

ActionExecutorOutput

Represents the output of the Action Executor.

ActionOutput

Represents the result of executing an action.

AddToFolderAction

Represents the action responsible for adding document under a folder.

CloudAiDocumentOption

Request Option for processing Cloud AI Document in CW Document.

CreateDocumentLinkRequest

Request message for DocumentLinkService.CreateDocumentLink.

CreateDocumentMetadata

Metadata object for CreateDocument request (currently empty).

CreateDocumentRequest

Request message for DocumentService.CreateDocument.

CreateDocumentResponse

Response message for DocumentService.CreateDocument.

CreateDocumentSchemaRequest

Request message for DocumentSchemaService.CreateDocumentSchema.

CreateRuleSetRequest

Request message for RuleSetService.CreateRuleSet.

CreateSynonymSetRequest

Request message for SynonymSetService.CreateSynonymSet.

CustomWeightsMetadata

To support the custom weighting across document schemas.

DataUpdateAction

Represents the action responsible for properties update operations.

DataValidationAction

Represents the action responsible for data validation operations.

DateTimeArray

DateTime values.

DateTimeTypeOptions

Configurations for a date time property.

DeleteDocumentAction

Represents the action responsible for deleting the document.

DeleteDocumentLinkRequest

Request message for DocumentLinkService.DeleteDocumentLink.

DeleteDocumentRequest

Request message for DocumentService.DeleteDocument.

DeleteDocumentSchemaRequest

Request message for DocumentSchemaService.DeleteDocumentSchema.

DeleteRuleSetRequest

Request message for RuleSetService.DeleteRuleSet.

DeleteSynonymSetRequest

Request message for SynonymSetService.DeleteSynonymSet.

Document

Defines the structure for content warehouse document proto.

DocumentLink

A document-link between source and target document.

DocumentQuery

DocumentReference

References to the documents.

DocumentSchema

A document schema used to define document structure.

EnumArray

Enum values.

EnumTypeOptions

Configurations for an enum/categorical property.

EnumValue

Represents the string value of the enum field.

ExportToCdwPipeline

The configuration of exporting documents from the Document Warehouse to CDW pipeline.

FetchAclRequest

Request message for DocumentService.FetchAcl

FetchAclResponse

Response message for DocumentService.FetchAcl.

FileTypeFilter

Filter for the specific types of documents returned.

FloatArray

Float values.

FloatTypeOptions

Configurations for a float property.

GcsIngestPipeline

The configuration of the Cloud Storage Ingestion pipeline.

GcsIngestWithDocAiProcessorsPipeline

The configuration of the Cloud Storage Ingestion with DocAI Processors pipeline.

GetDocumentRequest

Request message for DocumentService.GetDocument.

GetDocumentSchemaRequest

Request message for DocumentSchemaService.GetDocumentSchema.

GetRuleSetRequest

Request message for RuleSetService.GetRuleSet.

GetSynonymSetRequest

Request message for SynonymSetService.GetSynonymSet. Will return synonymSet for a certain context.

HistogramQuery

The histogram request.

HistogramQueryPropertyNameFilter

HistogramQueryResult

Histogram result that matches [HistogramQuery][google.cloud.contentwarehouse.v1.HistogramQuery] specified in searches.

IngestPipelineConfig

The ingestion pipeline config.

IntegerArray

Integer values.

IntegerTypeOptions

Configurations for an integer property.

InvalidRule

A triggered rule that failed the validation check(s) after parsing.

ListDocumentSchemasRequest

Request message for DocumentSchemaService.ListDocumentSchemas.

ListDocumentSchemasResponse

Response message for DocumentSchemaService.ListDocumentSchemas.

ListLinkedSourcesRequest

Response message for DocumentLinkService.ListLinkedSources.

ListLinkedSourcesResponse

Response message for DocumentLinkService.ListLinkedSources.

ListLinkedTargetsRequest

Request message for DocumentLinkService.ListLinkedTargets.

ListLinkedTargetsResponse

Response message for DocumentLinkService.ListLinkedTargets.

ListRuleSetsRequest

Request message for RuleSetService.ListRuleSets.

ListRuleSetsResponse

Response message for RuleSetService.ListRuleSets.

ListSynonymSetsRequest

Request message for SynonymSetService.ListSynonymSets. Will return all synonymSets belonging to the customer project.

ListSynonymSetsResponse

Response message for SynonymSetService.ListSynonymSets.

LockDocumentRequest

Request message for DocumentService.LockDocument.

MapProperty

Map property value. Represents a structured entries of key value pairs, consisting of field names which map to dynamically typed values.

MapTypeOptions

Configurations for a Map property.

MergeFieldsOptions

Options for merging updated fields.

ProcessWithDocAiPipeline

The configuration of processing documents in Document Warehouse with DocAi processors pipeline.

ProcessorInfo

The DocAI processor information.

Property

Property of a document.

PropertyArray

Property values.

PropertyDefinition

Defines the metadata for a schema property.

PropertyFilter

PropertyTypeOptions

Configurations for a nested structured data property.

PublishAction

Represents the action responsible for publishing messages to a Pub/Sub topic.

QaResult

Additional result info for the question-answering feature.

RemoveFromFolderAction

Represents the action responsible for remove a document from a specific folder.

RequestMetadata

Meta information is used to improve the performance of the service.

ResponseMetadata

Additional information returned to client, such as debugging information.

Rule

Represents the rule for a content warehouse trigger.

RuleActionsPair

Represents a rule and outputs of associated actions.

RuleEngineOutput

Records the output of Rule Engine including rule evaluation and actions result.

RuleEvaluatorOutput

Represents the output of the Rule Evaluator.

RuleSet

Represents a set of rules from a single customer.

RunPipelineMetadata

Metadata message of RunPipeline method.

RunPipelineRequest

Request message for DocumentService.RunPipeline.

RunPipelineResponse

Response message of RunPipeline method.

SearchDocumentsRequest

Request message for DocumentService.SearchDocuments.

SearchDocumentsResponse

Response message for DocumentService.SearchDocuments.

SetAclRequest

Request message for DocumentService.SetAcl.

SetAclResponse

Response message for DocumentService.SetAcl.

SynonymSet

Represents a list of synonyms for a given context. For example a context “sales” could contain: Synonym 1: sale, invoice, bill, order Synonym 2: money, credit, finance, payment Synonym 3: shipping, freight, transport Each SynonymSets should be disjoint

TextArray

String/text values.

TextTypeOptions

Configurations for a text property.

TimeFilter

Filter on create timestamp or update timestamp of documents.

TimestampArray

Timestamp values.

TimestampTypeOptions

Configurations for a timestamp property.

TimestampValue

Timestamp value type.

UpdateDocumentMetadata

Metadata object for UpdateDocument request (currently empty).

UpdateDocumentRequest

Request message for DocumentService.UpdateDocument.

UpdateDocumentResponse

Response message for DocumentService.UpdateDocument.

UpdateDocumentSchemaRequest

Request message for DocumentSchemaService.UpdateDocumentSchema.

UpdateOptions

Options for Update operations.

UpdateRuleSetRequest

Request message for RuleSetService.UpdateRuleSet.

UpdateSynonymSetRequest

Request message for SynonymSetService.UpdateSynonymSet. Removes the SynonymSet for the specified context and replaces it with the SynonymSet in this request.

UserInfo

The user information.

Value

Value represents a dynamically typed value which can be either be a float, a integer, a string, or a datetime value. A producer of value is expected to set one of these variants. Absence of any variant indicates an error.

WeightedSchemaProperty

Specifies the schema property name.

Enums

AccessControlMode

Access Control Mode.

ContentCategory

When a raw document or structured content is supplied, this stores the content category.

DatabaseType

Type of database used by the customer

DocumentCreatorDefaultRole

The default role of the document creator.

RawDocumentFileType

When a raw document is supplied, this indicates the file format

UpdateType

Update type of the requests.