Module google_api_proto::google::cloud::datacatalog::lineage::v1

Modules

lineage_client

Generated client implementations.

operation_metadata

Nested message and enum types in OperationMetadata.

origin

Nested message and enum types in Origin.

run

Nested message and enum types in Run.

search_links_request

Nested message and enum types in SearchLinksRequest.

Structs

BatchSearchLinkProcessesRequest

Request message for [BatchSearchLinkProcesses][google.cloud.datacatalog.lineage.v1.Lineage.BatchSearchLinkProcesses].

BatchSearchLinkProcessesResponse

Response message for [BatchSearchLinkProcesses][google.cloud.datacatalog.lineage.v1.Lineage.BatchSearchLinkProcesses].

CreateLineageEventRequest

Request message for [CreateLineageEvent][google.cloud.datacatalog.lineage.v1.CreateLineageEvent].

CreateProcessRequest

Request message for [CreateProcess][google.cloud.datacatalog.lineage.v1.CreateProcess].

CreateRunRequest

Request message for [CreateRun][google.cloud.datacatalog.lineage.v1.CreateRun].

DeleteLineageEventRequest

Request message for [DeleteLineageEvent][google.cloud.datacatalog.lineage.v1.DeleteLineageEvent].

DeleteProcessRequest

Request message for [DeleteProcess][google.cloud.datacatalog.lineage.v1.DeleteProcess].

DeleteRunRequest

Request message for [DeleteRun][google.cloud.datacatalog.lineage.v1.DeleteRun].

EntityReference

The soft reference to everything you can attach a lineage event to.

EventLink

A lineage between source and target entities.

GetLineageEventRequest

Request message for [GetLineageEvent][google.cloud.datacatalog.lineage.v1.GetLineageEvent].

GetProcessRequest

Request message for [GetProcess][google.cloud.datacatalog.lineage.v1.GetProcess].

GetRunRequest

Request message for [GetRun][google.cloud.datacatalog.lineage.v1.GetRun].

LineageEvent

A lineage event represents an operation on assets. Within the operation, the data flows from the source to the target defined in the links field.

Link

Links represent the data flow between source (upstream) and target (downstream) assets in transformation pipelines.

ListLineageEventsRequest

Request message for [ListLineageEvents][google.cloud.datacatalog.lineage.v1.ListLineageEvents].

ListLineageEventsResponse

Response message for [ListLineageEvents][google.cloud.datacatalog.lineage.v1.ListLineageEvents].

ListProcessesRequest

Request message for [ListProcesses][google.cloud.datacatalog.lineage.v1.ListProcesses].

ListProcessesResponse

Response message for [ListProcesses][google.cloud.datacatalog.lineage.v1.ListProcesses].

ListRunsRequest

Request message for [ListRuns][google.cloud.datacatalog.lineage.v1.ListRuns].

ListRunsResponse

Response message for [ListRuns][google.cloud.datacatalog.lineage.v1.ListRuns].

OperationMetadata

Metadata describing the operation.

Origin

Origin of a process.

Process

A process is the definition of a data transformation operation.

ProcessLinkInfo

Link details.

ProcessLinks

Links associated with a specific process.

ProcessOpenLineageRunEventRequest

Request message for [ProcessOpenLineageRunEvent][google.cloud.datacatalog.lineage.v1.ProcessOpenLineageRunEvent].

ProcessOpenLineageRunEventResponse

Response message for [ProcessOpenLineageRunEvent][google.cloud.datacatalog.lineage.v1.ProcessOpenLineageRunEvent].

Run

A lineage run represents an execution of a process that creates lineage events.

SearchLinksRequest

Request message for [SearchLinks][google.cloud.datacatalog.lineage.v1.Lineage.SearchLinks].

SearchLinksResponse

Response message for [SearchLinks][google.cloud.datacatalog.lineage.v1.Lineage.SearchLinks].

UpdateProcessRequest

Request message for [UpdateProcess][google.cloud.datacatalog.lineage.v1.UpdateProcess].

UpdateRunRequest

Request message for [UpdateRun][google.cloud.datacatalog.lineage.v1.UpdateRun].