Module google_api_proto::google::cloud::dataform::v1alpha2

Modules

compilation_result

Nested message and enum types in CompilationResult.

compilation_result_action

Nested message and enum types in CompilationResultAction.

dataform_client

Generated client implementations.

fetch_file_git_statuses_response

Nested message and enum types in FetchFileGitStatusesResponse.

query_directory_contents_response

Nested message and enum types in QueryDirectoryContentsResponse.

relation_descriptor

Nested message and enum types in RelationDescriptor.

repository

Nested message and enum types in Repository.

workflow_invocation

Nested message and enum types in WorkflowInvocation.

workflow_invocation_action

Nested message and enum types in WorkflowInvocationAction.

Structs

CancelWorkflowInvocationRequest

CancelWorkflowInvocation request message.

CommitAuthor

Represents the author of a Git commit.

CommitWorkspaceChangesRequest

CommitWorkspaceChanges request message.

CompilationResult

Represents the result of compiling a Dataform project.

CompilationResultAction

Represents a single Dataform action in a compilation result.

CreateCompilationResultRequest

CreateCompilationResult request message.

CreateRepositoryRequest

CreateRepository request message.

CreateWorkflowInvocationRequest

CreateWorkflowInvocation request message.

CreateWorkspaceRequest

CreateWorkspace request message.

DeleteRepositoryRequest

DeleteRepository request message.

DeleteWorkflowInvocationRequest

DeleteWorkflowInvocation request message.

DeleteWorkspaceRequest

DeleteWorkspace request message.

FetchFileDiffRequest

FetchFileDiff request message.

FetchFileDiffResponse

FetchFileDiff response message.

FetchFileGitStatusesRequest

FetchFileGitStatuses request message.

FetchFileGitStatusesResponse

FetchFileGitStatuses response message.

FetchGitAheadBehindRequest

FetchGitAheadBehind request message.

FetchGitAheadBehindResponse

FetchGitAheadBehind response message.

FetchRemoteBranchesRequest

FetchRemoteBranches request message.

FetchRemoteBranchesResponse

FetchRemoteBranches response message.

GetCompilationResultRequest

GetCompilationResult request message.

GetRepositoryRequest

GetRepository request message.

GetWorkflowInvocationRequest

GetWorkflowInvocation request message.

GetWorkspaceRequest

GetWorkspace request message.

InstallNpmPackagesRequest

InstallNpmPackages request message.

InstallNpmPackagesResponse

InstallNpmPackages response message.

ListCompilationResultsRequest

ListCompilationResults request message.

ListCompilationResultsResponse

ListCompilationResults response message.

ListRepositoriesRequest

ListRepositories request message.

ListRepositoriesResponse

ListRepositories response message.

ListWorkflowInvocationsRequest

ListWorkflowInvocations request message.

ListWorkflowInvocationsResponse

ListWorkflowInvocations response message.

ListWorkspacesRequest

ListWorkspaces request message.

ListWorkspacesResponse

ListWorkspaces response message.

MakeDirectoryRequest

MakeDirectory request message.

MakeDirectoryResponse

MakeDirectory response message.

MoveDirectoryRequest

MoveDirectory request message.

MoveDirectoryResponse

MoveDirectory response message.

MoveFileRequest

MoveFile request message.

MoveFileResponse

MoveFile response message.

PullGitCommitsRequest

PullGitCommits request message.

PushGitCommitsRequest

PushGitCommits request message.

QueryCompilationResultActionsRequest

QueryCompilationResultActions request message.

QueryCompilationResultActionsResponse

QueryCompilationResultActions response message.

QueryDirectoryContentsRequest

QueryDirectoryContents request message.

QueryDirectoryContentsResponse

QueryDirectoryContents response message.

QueryWorkflowInvocationActionsRequest

QueryWorkflowInvocationActions request message.

QueryWorkflowInvocationActionsResponse

QueryWorkflowInvocationActions response message.

ReadFileRequest

ReadFile request message.

ReadFileResponse

ReadFile response message.

RelationDescriptor

Describes a relation and its columns.

RemoveDirectoryRequest

RemoveDirectory request message.

RemoveFileRequest

RemoveFile request message.

Repository

Represents a Dataform Git repository.

ResetWorkspaceChangesRequest

ResetWorkspaceChanges request message.

Target

Represents an action identifier. If the action writes output, the output will be written to the referenced database object.

UpdateRepositoryRequest

UpdateRepository request message.

WorkflowInvocation

Represents a single invocation of a compilation result.

WorkflowInvocationAction

Represents a single action in a workflow invocation.

Workspace

Represents a Dataform Git workspace.

WriteFileRequest

WriteFile request message.

WriteFileResponse

WriteFile response message.