Module google_api_proto::google::cloud::dataform::v1beta1

Modules

commit_repository_changes_request

Nested message and enum types in CommitRepositoryChangesRequest.

compilation_result

Nested message and enum types in CompilationResult.

compilation_result_action

Nested message and enum types in CompilationResultAction.

compute_repository_access_token_status_response

Nested message and enum types in ComputeRepositoryAccessTokenStatusResponse.

dataform_client

Generated client implementations.

directory_entry

Nested message and enum types in DirectoryEntry.

fetch_file_git_statuses_response

Nested message and enum types in FetchFileGitStatusesResponse.

relation_descriptor

Nested message and enum types in RelationDescriptor.

release_config

Nested message and enum types in ReleaseConfig.

repository

Nested message and enum types in Repository.

workflow_config

Nested message and enum types in WorkflowConfig.

workflow_invocation

Nested message and enum types in WorkflowInvocation.

workflow_invocation_action

Nested message and enum types in WorkflowInvocationAction.

Structs

CancelWorkflowInvocationRequest

CancelWorkflowInvocation request message.

CodeCompilationConfig

Configures various aspects of Dataform code compilation.

CommitAuthor

Represents the author of a Git commit.

CommitLogEntry

Represents a single commit log.

CommitMetadata

Represents a Dataform Git commit.

CommitRepositoryChangesRequest

CommitRepositoryChanges request message.

CommitWorkspaceChangesRequest

CommitWorkspaceChanges request message.

CompilationResult

Represents the result of compiling a Dataform project.

CompilationResultAction

Represents a single Dataform action in a compilation result.

ComputeRepositoryAccessTokenStatusRequest

ComputeRepositoryAccessTokenStatus request message.

ComputeRepositoryAccessTokenStatusResponse

ComputeRepositoryAccessTokenStatus response message.

CreateCompilationResultRequest

CreateCompilationResult request message.

CreateReleaseConfigRequest

CreateReleaseConfig request message.

CreateRepositoryRequest

CreateRepository request message.

CreateWorkflowConfigRequest

CreateWorkflowConfig request message.

CreateWorkflowInvocationRequest

CreateWorkflowInvocation request message.

CreateWorkspaceRequest

CreateWorkspace request message.

DeleteReleaseConfigRequest

DeleteReleaseConfig request message.

DeleteRepositoryRequest

DeleteRepository request message.

DeleteWorkflowConfigRequest

DeleteWorkflowConfig request message.

DeleteWorkflowInvocationRequest

DeleteWorkflowInvocation request message.

DeleteWorkspaceRequest

DeleteWorkspace request message.

DirectoryEntry

Represents a single entry in a directory.

FetchFileDiffRequest

FetchFileDiff request message.

FetchFileDiffResponse

FetchFileDiff response message.

FetchFileGitStatusesRequest

FetchFileGitStatuses request message.

FetchFileGitStatusesResponse

FetchFileGitStatuses response message.

FetchGitAheadBehindRequest

FetchGitAheadBehind request message.

FetchGitAheadBehindResponse

FetchGitAheadBehind response message.

FetchRemoteBranchesRequest

FetchRemoteBranches request message.

FetchRemoteBranchesResponse

FetchRemoteBranches response message.

FetchRepositoryHistoryRequest

FetchRepositoryHistory request message.

FetchRepositoryHistoryResponse

FetchRepositoryHistory response message.

GetCompilationResultRequest

GetCompilationResult request message.

GetReleaseConfigRequest

GetReleaseConfig request message.

GetRepositoryRequest

GetRepository request message.

GetWorkflowConfigRequest

GetWorkflowConfig request message.

GetWorkflowInvocationRequest

GetWorkflowInvocation request message.

GetWorkspaceRequest

GetWorkspace request message.

InstallNpmPackagesRequest

InstallNpmPackages request message.

InstallNpmPackagesResponse

InstallNpmPackages response message.

InvocationConfig

Includes various configuration options for a workflow invocation. If both included_targets and included_tags are unset, all actions will be included.

ListCompilationResultsRequest

ListCompilationResults request message.

ListCompilationResultsResponse

ListCompilationResults response message.

ListReleaseConfigsRequest

ListReleaseConfigs request message.

ListReleaseConfigsResponse

ListReleaseConfigs response message.

ListRepositoriesRequest

ListRepositories request message.

ListRepositoriesResponse

ListRepositories response message.

ListWorkflowConfigsRequest

ListWorkflowConfigs request message.

ListWorkflowConfigsResponse

ListWorkflowConfigs response message.

ListWorkflowInvocationsRequest

ListWorkflowInvocations request message.

ListWorkflowInvocationsResponse

ListWorkflowInvocations response message.

ListWorkspacesRequest

ListWorkspaces request message.

ListWorkspacesResponse

ListWorkspaces response message.

MakeDirectoryRequest

MakeDirectory request message.

MakeDirectoryResponse

MakeDirectory response message.

MoveDirectoryRequest

MoveDirectory request message.

MoveDirectoryResponse

MoveDirectory response message.

MoveFileRequest

MoveFile request message.

MoveFileResponse

MoveFile response message.

PullGitCommitsRequest

PullGitCommits request message.

PushGitCommitsRequest

PushGitCommits request message.

QueryCompilationResultActionsRequest

QueryCompilationResultActions request message.

QueryCompilationResultActionsResponse

QueryCompilationResultActions response message.

QueryDirectoryContentsRequest

QueryDirectoryContents request message.

QueryDirectoryContentsResponse

QueryDirectoryContents response message.

QueryRepositoryDirectoryContentsRequest

QueryRepositoryDirectoryContents request message.

QueryRepositoryDirectoryContentsResponse

QueryRepositoryDirectoryContents response message.

QueryWorkflowInvocationActionsRequest

QueryWorkflowInvocationActions request message.

QueryWorkflowInvocationActionsResponse

QueryWorkflowInvocationActions response message.

ReadFileRequest

ReadFile request message.

ReadFileResponse

ReadFile response message.

ReadRepositoryFileRequest

ReadRepositoryFile request message.

ReadRepositoryFileResponse

ReadRepositoryFile response message.

RelationDescriptor

Describes a relation and its columns.

ReleaseConfig

Represents a Dataform release configuration.

RemoveDirectoryRequest

RemoveDirectory request message.

RemoveFileRequest

RemoveFile request message.

Repository

Represents a Dataform Git repository.

ResetWorkspaceChangesRequest

ResetWorkspaceChanges request message.

Target

Represents an action identifier. If the action writes output, the output will be written to the referenced database object.

UpdateReleaseConfigRequest

UpdateReleaseConfig request message.

UpdateRepositoryRequest

UpdateRepository request message.

UpdateWorkflowConfigRequest

UpdateWorkflowConfig request message.

WorkflowConfig

Represents a Dataform workflow configuration.

WorkflowInvocation

Represents a single invocation of a compilation result.

WorkflowInvocationAction

Represents a single action in a workflow invocation.

Workspace

Represents a Dataform Git workspace.

WriteFileRequest

WriteFile request message.

WriteFileResponse

WriteFile response message.