Module google_api_proto::google::cloud::datafusion::v1beta1

Modules

accelerator

Nested message and enum types in Accelerator.

data_fusion_client

Generated client implementations.

instance

Nested message and enum types in Instance.

version

Nested message and enum types in Version.

Structs

Accelerator

Identifies Data Fusion accelerators for an instance.

AddDnsPeeringRequest

Request message to create dns peering.

AddDnsPeeringResponse

Response message for set dns peering method.

CreateInstanceRequest

Request message for creating a Data Fusion instance.

CryptoKeyConfig

The crypto key configuration. This field is used by the Customer-managed encryption keys (CMEK) feature.

DeleteInstanceRequest

Request message for deleting a Data Fusion instance.

DnsPeering

DNS peering configuration. These configurations are used to create DNS peering with the customer Cloud DNS.

GetInstanceRequest

Request message for getting details about a Data Fusion instance.

IamPolicy

IAMPolicy encapsulates the IAM policy name, definition and status of policy fetching.

Instance

Represents a Data Fusion instance.

ListAvailableVersionsRequest

Request message for the list available versions request.

ListAvailableVersionsResponse

Response message for the list available versions request.

ListDnsPeeringsRequest

List dns peering request.

ListDnsPeeringsResponse

List dns peering response.

ListInstancesRequest

Request message for listing Data Fusion instances.

ListInstancesResponse

Response message for the list instance request.

ListNamespacesRequest

List namespaces request.

ListNamespacesResponse

List namespaces response.

Namespace

Represents the information of a namespace

NetworkConfig

Network configuration for a Data Fusion instance. These configurations are used for peering with the customer network. Configurations are optional when a public Data Fusion instance is to be created. However, providing these configurations allows several benefits, such as reduced network latency while accessing the customer resources from managed Data Fusion instance nodes, as well as access to the customer on-prem resources.

OperationMetadata

Represents the metadata of a long-running operation.

RemoveDnsPeeringRequest

Request message to remove dns peering.

RemoveDnsPeeringResponse

Response message for set dns peering method.

RemoveIamPolicyRequest

Request message for RemoveIamPolicy method.

RemoveIamPolicyResponse

Response message for RemoveIamPolicy method.

RestartInstanceRequest

Request message for restarting a Data Fusion instance.

UpdateInstanceRequest

Request message for updating a Data Fusion instance. Data Fusion only allows updating the labels, options, and stack driver settings.

UpgradeInstanceRequest

Request message for upgrading a Data Fusion instance. To change the instance properties, instance update should be used.

Version

The Data Fusion version.

Enums

NamespaceView

A view for Namespace