Struct google_api_proto::google::cloud::datalabeling::v1beta1::data_labeling_service_client::DataLabelingServiceClient

pub struct DataLabelingServiceClient<T> { /* private fields */ }

Service for the AI Platform Data Labeling API.

Implementations

impl<T> DataLabelingServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> DataLabelingServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_dataset(
&mut self,
request: impl IntoRequest<CreateDatasetRequest>
) -> Result<Response<Dataset>, Status>

Creates dataset. If success return a Dataset resource.

pub async fn get_dataset(
&mut self,
request: impl IntoRequest<GetDatasetRequest>
) -> Result<Response<Dataset>, Status>

Gets dataset by resource name.

pub async fn list_datasets(
&mut self,
request: impl IntoRequest<ListDatasetsRequest>
) -> Result<Response<ListDatasetsResponse>, Status>

Lists datasets under a project. Pagination is supported.

pub async fn delete_dataset(
&mut self,
request: impl IntoRequest<DeleteDatasetRequest>
) -> Result<Response<()>, Status>

Deletes a dataset by resource name.

pub async fn import_data(
&mut self,
request: impl IntoRequest<ImportDataRequest>
) -> Result<Response<Operation>, Status>

Imports data into dataset based on source locations defined in request. It can be called multiple times for the same dataset. Each dataset can only have one long running operation running on it. For example, no labeling task (also long running operation) can be started while importing is still ongoing. Vice versa.

pub async fn export_data(
&mut self,
request: impl IntoRequest<ExportDataRequest>
) -> Result<Response<Operation>, Status>

Exports data and annotations from dataset.

pub async fn get_data_item(
&mut self,
request: impl IntoRequest<GetDataItemRequest>
) -> Result<Response<DataItem>, Status>

Gets a data item in a dataset by resource name. This API can be called after data are imported into dataset.

pub async fn list_data_items(
&mut self,
request: impl IntoRequest<ListDataItemsRequest>
) -> Result<Response<ListDataItemsResponse>, Status>

Lists data items in a dataset. This API can be called after data are imported into dataset. Pagination is supported.

pub async fn get_annotated_dataset(
&mut self,
request: impl IntoRequest<GetAnnotatedDatasetRequest>
) -> Result<Response<AnnotatedDataset>, Status>

Gets an annotated dataset by resource name.

pub async fn list_annotated_datasets(
&mut self,
request: impl IntoRequest<ListAnnotatedDatasetsRequest>
) -> Result<Response<ListAnnotatedDatasetsResponse>, Status>

Lists annotated datasets for a dataset. Pagination is supported.

pub async fn delete_annotated_dataset(
&mut self,
request: impl IntoRequest<DeleteAnnotatedDatasetRequest>
) -> Result<Response<()>, Status>

Deletes an annotated dataset by resource name.

pub async fn label_image(
&mut self,
request: impl IntoRequest<LabelImageRequest>
) -> Result<Response<Operation>, Status>

Starts a labeling task for image. The type of image labeling task is configured by feature in the request.

pub async fn label_video(
&mut self,
request: impl IntoRequest<LabelVideoRequest>
) -> Result<Response<Operation>, Status>

Starts a labeling task for video. The type of video labeling task is configured by feature in the request.

pub async fn label_text(
&mut self,
request: impl IntoRequest<LabelTextRequest>
) -> Result<Response<Operation>, Status>

Starts a labeling task for text. The type of text labeling task is configured by feature in the request.

pub async fn get_example(
&mut self,
request: impl IntoRequest<GetExampleRequest>
) -> Result<Response<Example>, Status>

Gets an example by resource name, including both data and annotation.

pub async fn list_examples(
&mut self,
request: impl IntoRequest<ListExamplesRequest>
) -> Result<Response<ListExamplesResponse>, Status>

Lists examples in an annotated dataset. Pagination is supported.

pub async fn create_annotation_spec_set(
&mut self,
request: impl IntoRequest<CreateAnnotationSpecSetRequest>
) -> Result<Response<AnnotationSpecSet>, Status>

Creates an annotation spec set by providing a set of labels.

pub async fn get_annotation_spec_set(
&mut self,
request: impl IntoRequest<GetAnnotationSpecSetRequest>
) -> Result<Response<AnnotationSpecSet>, Status>

Gets an annotation spec set by resource name.

pub async fn list_annotation_spec_sets(
&mut self,
request: impl IntoRequest<ListAnnotationSpecSetsRequest>
) -> Result<Response<ListAnnotationSpecSetsResponse>, Status>

Lists annotation spec sets for a project. Pagination is supported.

pub async fn delete_annotation_spec_set(
&mut self,
request: impl IntoRequest<DeleteAnnotationSpecSetRequest>
) -> Result<Response<()>, Status>

Deletes an annotation spec set by resource name.

pub async fn create_instruction(
&mut self,
request: impl IntoRequest<CreateInstructionRequest>
) -> Result<Response<Operation>, Status>

Creates an instruction for how data should be labeled.

pub async fn get_instruction(
&mut self,
request: impl IntoRequest<GetInstructionRequest>
) -> Result<Response<Instruction>, Status>

Gets an instruction by resource name.

pub async fn list_instructions(
&mut self,
request: impl IntoRequest<ListInstructionsRequest>
) -> Result<Response<ListInstructionsResponse>, Status>

Lists instructions for a project. Pagination is supported.

pub async fn delete_instruction(
&mut self,
request: impl IntoRequest<DeleteInstructionRequest>
) -> Result<Response<()>, Status>

Deletes an instruction object by resource name.

pub async fn get_evaluation(
&mut self,
request: impl IntoRequest<GetEvaluationRequest>
) -> Result<Response<Evaluation>, Status>

Gets an evaluation by resource name (to search, use [projects.evaluations.search][google.cloud.datalabeling.v1beta1.DataLabelingService.SearchEvaluations]).

pub async fn search_evaluations(
&mut self,
request: impl IntoRequest<SearchEvaluationsRequest>
) -> Result<Response<SearchEvaluationsResponse>, Status>

Searches [evaluations][google.cloud.datalabeling.v1beta1.Evaluation] within a project.

pub async fn search_example_comparisons(
&mut self,
request: impl IntoRequest<SearchExampleComparisonsRequest>
) -> Result<Response<SearchExampleComparisonsResponse>, Status>

Searches example comparisons from an evaluation. The return format is a list of example comparisons that show ground truth and prediction(s) for a single input. Search by providing an evaluation ID.

pub async fn create_evaluation_job(
&mut self,
request: impl IntoRequest<CreateEvaluationJobRequest>
) -> Result<Response<EvaluationJob>, Status>

Creates an evaluation job.

pub async fn update_evaluation_job(
&mut self,
request: impl IntoRequest<UpdateEvaluationJobRequest>
) -> Result<Response<EvaluationJob>, Status>

Updates an evaluation job. You can only update certain fields of the job’s [EvaluationJobConfig][google.cloud.datalabeling.v1beta1.EvaluationJobConfig]: humanAnnotationConfig.instruction, exampleCount, and exampleSamplePercentage.

If you want to change any other aspect of the evaluation job, you must delete the job and create a new one.

pub async fn get_evaluation_job(
&mut self,
request: impl IntoRequest<GetEvaluationJobRequest>
) -> Result<Response<EvaluationJob>, Status>

Gets an evaluation job by resource name.

pub async fn pause_evaluation_job(
&mut self,
request: impl IntoRequest<PauseEvaluationJobRequest>
) -> Result<Response<()>, Status>

Pauses an evaluation job. Pausing an evaluation job that is already in a PAUSED state is a no-op.

pub async fn resume_evaluation_job(
&mut self,
request: impl IntoRequest<ResumeEvaluationJobRequest>
) -> Result<Response<()>, Status>

Resumes a paused evaluation job. A deleted evaluation job can’t be resumed. Resuming a running or scheduled evaluation job is a no-op.

pub async fn delete_evaluation_job(
&mut self,
request: impl IntoRequest<DeleteEvaluationJobRequest>
) -> Result<Response<()>, Status>

Stops and deletes an evaluation job.

pub async fn list_evaluation_jobs(
&mut self,
request: impl IntoRequest<ListEvaluationJobsRequest>
) -> Result<Response<ListEvaluationJobsResponse>, Status>

Lists all evaluation jobs within a project with possible filters. Pagination is supported.