Module google_api_proto::google::cloud::dataplex::v1

Modules

action

Nested message and enum types in Action.

asset

Nested message and enum types in Asset.

content

Nested message and enum types in Content.

content_service_client

Generated client implementations.

dataplex_service_client

Generated client implementations.

discovery_event

Nested message and enum types in DiscoveryEvent.

entity

Nested message and enum types in Entity.

environment

Nested message and enum types in Environment.

get_content_request

Nested message and enum types in GetContentRequest.

get_entity_request

Nested message and enum types in GetEntityRequest.

job

Nested message and enum types in Job.

job_event

Nested message and enum types in JobEvent.

lake

Nested message and enum types in Lake.

list_entities_request

Nested message and enum types in ListEntitiesRequest.

metadata_service_client

Generated client implementations.

schema

Nested message and enum types in Schema.

session_event

Nested message and enum types in SessionEvent.

storage_format

Nested message and enum types in StorageFormat.

task

Nested message and enum types in Task.

zone

Nested message and enum types in Zone.

Structs

Action

Action represents an issue requiring administrator action for resolution.

Asset

An asset represents a cloud resource that is being managed within a lake as a member of a zone.

AssetStatus

Aggregated status of the underlying assets of a lake or zone.

CancelJobRequest

Cancel task jobs.

Content

Content represents a user-visible notebook or a sql script

CreateAssetRequest

Create asset request.

CreateContentRequest

Create content request.

CreateEntityRequest

Create a metadata entity request.

CreateEnvironmentRequest

Create environment request.

CreateLakeRequest

Create lake request.

CreatePartitionRequest

Create metadata partition request.

CreateTaskRequest

Create task request.

CreateZoneRequest

Create zone request.

DeleteAssetRequest

Delete asset request.

DeleteContentRequest

Delete content request.

DeleteEntityRequest

Delete a metadata entity request.

DeleteEnvironmentRequest

Delete environment request.

DeleteLakeRequest

Delete lake request.

DeletePartitionRequest

Delete metadata partition request.

DeleteTaskRequest

Delete task request.

DeleteZoneRequest

Delete zone request.

DiscoveryEvent

The payload associated with Discovery data processing.

Entity

Represents tables and fileset metadata contained within a zone.

Environment

Environment represents a user-visible compute infrastructure for analytics within a lake.

GetAssetRequest

Get asset request.

GetContentRequest

Get content request.

GetEntityRequest

Get metadata entity request.

GetEnvironmentRequest

Get environment request.

GetJobRequest

Get job request.

GetLakeRequest

Get lake request.

GetPartitionRequest

Get metadata partition request.

GetTaskRequest

Get task request.

GetZoneRequest

Get zone request.

Job

A job represents an instance of a task.

JobEvent

The payload associated with Job logs that contains events describing jobs that have run within a Lake.

Lake

A lake is a centralized repository for managing enterprise data across the organization distributed across many cloud projects, and stored in a variety of storage services such as Google Cloud Storage and BigQuery. The resources attached to a lake are referred to as managed resources. Data within these managed resources can be structured or unstructured. A lake provides data admins with tools to organize, secure and manage their data at scale, and provides data scientists and data engineers an integrated experience to easily search, discover, analyze and transform data and associated metadata.

ListActionsResponse

List actions response.

ListAssetActionsRequest

List asset actions request.

ListAssetsRequest

List assets request.

ListAssetsResponse

List assets response.

ListContentRequest

List content request. Returns the BASIC Content view.

ListContentResponse

List content response.

ListEntitiesRequest

List metadata entities request.

ListEntitiesResponse

List metadata entities response.

ListEnvironmentsRequest

List environments request.

ListEnvironmentsResponse

List environments response.

ListJobsRequest

List jobs request.

ListJobsResponse

List jobs response.

ListLakeActionsRequest

List lake actions request.

ListLakesRequest

List lakes request.

ListLakesResponse

List lakes response.

ListPartitionsRequest

List metadata partitions request.

ListPartitionsResponse

List metadata partitions response.

ListSessionsRequest

List sessions request.

ListSessionsResponse

List sessions response.

ListTasksRequest

List tasks request.

ListTasksResponse

List tasks response.

ListZoneActionsRequest

List zone actions request.

ListZonesRequest

List zones request.

ListZonesResponse

List zones response.

OperationMetadata

Represents the metadata of a long-running operation.

Partition

Represents partition metadata contained within entity instances.

Schema

Schema information describing the structure and layout of the data.

Session

Represents an active analyze session running for a user.

SessionEvent

These messages contain information about sessions within an environment. The monitored resource is ‘Environment’.

StorageFormat

Describes the format of the data within its storage location.

Task

A task represents a user-visible job.

UpdateAssetRequest

Update asset request.

UpdateContentRequest

Update content request.

UpdateEntityRequest

Update a metadata entity request. The exiting entity will be fully replaced by the entity in the request. The entity ID is mutable. To modify the ID, use the current entity ID in the request URL and specify the new ID in the request body.

UpdateEnvironmentRequest

Update environment request.

UpdateLakeRequest

Update lake request.

UpdateTaskRequest

Update task request.

UpdateZoneRequest

Update zone request.

Zone

A zone represents a logical group of related assets within a lake. A zone can be used to map to organizational structure or represent stages of data readiness from raw to curated. It provides managing behavior that is shared or inherited by all contained assets.

Enums

State

State of a resource.

StorageSystem

Identifies the cloud system that manages the data storage.