Module google_api_proto::google::cloud::dataplex::v1::action

Nested message and enum types in Action.

Modules

incompatible_data_schema

Nested message and enum types in IncompatibleDataSchema.

invalid_data_partition

Nested message and enum types in InvalidDataPartition.

Structs

FailedSecurityPolicyApply

Failed to apply security policy to the managed resource(s) under a lake, zone or an asset. For a lake or zone resource, one or more underlying assets has a failure applying security policy to the associated managed resource.

IncompatibleDataSchema

Action details for incompatible schemas detected by discovery.

InvalidDataFormat

Action details for invalid or unsupported data files detected by discovery.

InvalidDataOrganization

Action details for invalid data arrangement.

InvalidDataPartition

Action details for invalid or unsupported partitions detected by discovery.

MissingData

Action details for absence of data detected by discovery.

MissingResource

Action details for resource references in assets that cannot be located.

UnauthorizedResource

Action details for unauthorized resource issues raised to indicate that the service account associated with the lake instance is not authorized to access or manage the resource associated with an asset.

Enums

Category

The category of issues.

Details

Additional details about the action based on the action category.