Struct google_api_proto::google::cloud::dataplex::v1::dataplex_service_client::DataplexServiceClient

pub struct DataplexServiceClient<T> { /* fields omitted */ }

Dataplex service provides data lakes as a service. The primary resources offered by this service are Lakes, Zones and Assets which collectively allow a data adminstrator to organize, manage, secure and catalog data across their organization located across cloud projects in a variety of storage systems including Cloud Storage and BigQuery.

Implementations

impl<T> DataplexServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> DataplexServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_lake(
&mut self,
request: impl IntoRequest<CreateLakeRequest>
) -> Result<Response<Operation>, Status>

Creates a lake resource.

pub async fn update_lake(
&mut self,
request: impl IntoRequest<UpdateLakeRequest>
) -> Result<Response<Operation>, Status>

Updates a lake resource.

pub async fn delete_lake(
&mut self,
request: impl IntoRequest<DeleteLakeRequest>
) -> Result<Response<Operation>, Status>

Deletes a lake resource. All zones within the lake must be deleted before the lake can be deleted.

pub async fn list_lakes(
&mut self,
request: impl IntoRequest<ListLakesRequest>
) -> Result<Response<ListLakesResponse>, Status>

Lists lake resources in a project and location.

pub async fn get_lake(
&mut self,
request: impl IntoRequest<GetLakeRequest>
) -> Result<Response<Lake>, Status>

Retrieves a lake resource.

pub async fn list_lake_actions(
&mut self,
request: impl IntoRequest<ListLakeActionsRequest>
) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in a lake.

pub async fn create_zone(
&mut self,
request: impl IntoRequest<CreateZoneRequest>
) -> Result<Response<Operation>, Status>

Creates a zone resource within a lake.

pub async fn update_zone(
&mut self,
request: impl IntoRequest<UpdateZoneRequest>
) -> Result<Response<Operation>, Status>

Updates a zone resource.

pub async fn delete_zone(
&mut self,
request: impl IntoRequest<DeleteZoneRequest>
) -> Result<Response<Operation>, Status>

Deletes a zone resource. All assets within a zone must be deleted before the zone can be deleted.

pub async fn list_zones(
&mut self,
request: impl IntoRequest<ListZonesRequest>
) -> Result<Response<ListZonesResponse>, Status>

Lists zone resources in a lake.

pub async fn get_zone(
&mut self,
request: impl IntoRequest<GetZoneRequest>
) -> Result<Response<Zone>, Status>

Retrieves a zone resource.

pub async fn list_zone_actions(
&mut self,
request: impl IntoRequest<ListZoneActionsRequest>
) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in a zone.

pub async fn create_asset(
&mut self,
request: impl IntoRequest<CreateAssetRequest>
) -> Result<Response<Operation>, Status>

Creates an asset resource.

pub async fn update_asset(
&mut self,
request: impl IntoRequest<UpdateAssetRequest>
) -> Result<Response<Operation>, Status>

Updates an asset resource.

pub async fn delete_asset(
&mut self,
request: impl IntoRequest<DeleteAssetRequest>
) -> Result<Response<Operation>, Status>

Deletes an asset resource. The referenced storage resource is detached (default) or deleted based on the associated Lifecycle policy.

pub async fn list_assets(
&mut self,
request: impl IntoRequest<ListAssetsRequest>
) -> Result<Response<ListAssetsResponse>, Status>

Lists asset resources in a zone.

pub async fn get_asset(
&mut self,
request: impl IntoRequest<GetAssetRequest>
) -> Result<Response<Asset>, Status>

Retrieves an asset resource.

pub async fn list_asset_actions(
&mut self,
request: impl IntoRequest<ListAssetActionsRequest>
) -> Result<Response<ListActionsResponse>, Status>

Lists action resources in an asset.

pub async fn create_task(
&mut self,
request: impl IntoRequest<CreateTaskRequest>
) -> Result<Response<Operation>, Status>

Creates a task resource within a lake.

pub async fn update_task(
&mut self,
request: impl IntoRequest<UpdateTaskRequest>
) -> Result<Response<Operation>, Status>

Update the task resource.

pub async fn delete_task(
&mut self,
request: impl IntoRequest<DeleteTaskRequest>
) -> Result<Response<Operation>, Status>

Delete the task resource.

pub async fn list_tasks(
&mut self,
request: impl IntoRequest<ListTasksRequest>
) -> Result<Response<ListTasksResponse>, Status>

Lists tasks under the given lake.

pub async fn get_task(
&mut self,
request: impl IntoRequest<GetTaskRequest>
) -> Result<Response<Task>, Status>

Get task resource.

pub async fn list_jobs(
&mut self,
request: impl IntoRequest<ListJobsRequest>
) -> Result<Response<ListJobsResponse>, Status>

Lists Jobs under the given task.

pub async fn get_job(
&mut self,
request: impl IntoRequest<GetJobRequest>
) -> Result<Response<Job>, Status>

Get job resource.

pub async fn cancel_job(
&mut self,
request: impl IntoRequest<CancelJobRequest>
) -> Result<Response<()>, Status>

Cancel jobs running for the task resource.