Module google_api_proto::google::cloud::datastream::v1

Modules

backfill_job

Nested message and enum types in BackfillJob.

connection_profile

Nested message and enum types in ConnectionProfile.

datastream_client

Generated client implementations.

destination_config

Nested message and enum types in DestinationConfig.

discover_connection_profile_request

Nested message and enum types in DiscoverConnectionProfileRequest.

discover_connection_profile_response

Nested message and enum types in DiscoverConnectionProfileResponse.

forward_ssh_tunnel_connectivity

Nested message and enum types in ForwardSshTunnelConnectivity.

gcs_destination_config

Nested message and enum types in GcsDestinationConfig.

json_file_format

Nested message and enum types in JsonFileFormat.

private_connection

Nested message and enum types in PrivateConnection.

source_config

Nested message and enum types in SourceConfig.

source_object_identifier

Nested message and enum types in SourceObjectIdentifier.

stream

Nested message and enum types in Stream.

validation

Nested message and enum types in Validation.

validation_message

Nested message and enum types in ValidationMessage.

Structs

AvroFileFormat

AVRO file format configuration.

BackfillJob

Represents a backfill job on a specific stream object.

ConnectionProfile

A set of reusable connection configurations to be used as a source or destination for a stream.

CreateConnectionProfileRequest

Request message for creating a connection profile.

CreatePrivateConnectionRequest

Request for creating a private connection.

CreateRouteRequest

Route creation request.

CreateStreamRequest

Request message for creating a stream.

DeleteConnectionProfileRequest

Request message for deleting a connection profile.

DeletePrivateConnectionRequest

Request to delete a private connection.

DeleteRouteRequest

Route deletion request.

DeleteStreamRequest

Request message for deleting a stream.

DestinationConfig

The configuration of the stream destination.

DiscoverConnectionProfileRequest

Request message for ‘discover’ ConnectionProfile request.

DiscoverConnectionProfileResponse

Response from a discover request.

Error

Represent a user-facing Error.

FetchStaticIpsRequest

Request message for ‘FetchStaticIps’ request.

FetchStaticIpsResponse

Response message for a ‘FetchStaticIps’ response.

ForwardSshTunnelConnectivity

Forward SSH Tunnel connectivity.

GcsDestinationConfig

Google Cloud Storage destination configuration

GcsProfile

Cloud Storage bucket profile.

GetConnectionProfileRequest

Request message for getting a connection profile.

GetPrivateConnectionRequest

Request to get a private connection configuration.

GetRouteRequest

Route get request.

GetStreamObjectRequest

Request for fetching a specific stream object.

GetStreamRequest

Request message for getting a stream.

JsonFileFormat

JSON file format configuration.

ListConnectionProfilesRequest

Request message for listing connection profiles.

ListConnectionProfilesResponse

Response message for listing connection profiles.

ListPrivateConnectionsRequest

Request for listing private connections.

ListPrivateConnectionsResponse

Response containing a list of private connection configurations.

ListRoutesRequest

Route list request.

ListRoutesResponse

Route list response.

ListStreamObjectsRequest

Request for listing all objects for a specific stream.

ListStreamObjectsResponse

Response containing the objects for a stream.

ListStreamsRequest

Request message for listing streams.

ListStreamsResponse

Response message for listing streams.

LookupStreamObjectRequest

Request for looking up a specific stream object by its source object identifier.

MysqlColumn

MySQL Column.

MysqlDatabase

MySQL database.

MysqlProfile

MySQL database profile.

MysqlRdbms

MySQL database structure

MysqlSourceConfig

MySQL source configuration

MysqlSslConfig

MySQL SSL configuration information.

MysqlTable

MySQL table.

OperationMetadata

Represents the metadata of the long-running operation.

OracleColumn

Oracle Column.

OracleProfile

Oracle database profile.

OracleRdbms

Oracle database structure.

OracleSchema

Oracle schema.

OracleSourceConfig

Oracle data source configuration

OracleTable

Oracle table.

PrivateConnection

The PrivateConnection resource is used to establish private connectivity between Datastream and a customer’s network.

PrivateConnectivity

Private Connectivity

Route

The route resource is the child of the private connection resource, used for defining a route for a private connection.

SourceConfig

The configuration of the stream source.

SourceObjectIdentifier

Represents an identifier of an object in the data source.

StartBackfillJobRequest

Request for manually initiating a backfill job for a specific stream object.

StartBackfillJobResponse

Response for manually initiating a backfill job for a specific stream object.

StaticServiceIpConnectivity

Static IP address connectivity.

StopBackfillJobRequest

Request for manually stopping a running backfill job for a specific stream object.

StopBackfillJobResponse

Response for manually stop a backfill job for a specific stream object.

Stream

A resource representing streaming data from a source to a destination.

StreamObject

A specific stream object (e.g a specific DB table).

UpdateConnectionProfileRequest

Connection profile update message.

UpdateStreamRequest

Request message for updating a stream.

Validation

A validation to perform on a stream.

ValidationMessage

Represent user-facing validation result message.

ValidationResult

Contains the current validation results.

VpcPeeringConfig

The VPC Peering configuration is used to create VPC peering between Datastream and the consumer’s VPC.