Struct google_api_proto::google::cloud::datastream::v1alpha1::datastream_client::DatastreamClient

pub struct DatastreamClient<T> { /* private fields */ }

Datastream service

Implementations

impl<T> DatastreamClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> DatastreamClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_connection_profiles( &mut self, request: impl IntoRequest<ListConnectionProfilesRequest> ) -> Result<Response<ListConnectionProfilesResponse>, Status>

Use this method to list connection profiles created in a project and location.

pub async fn get_connection_profile( &mut self, request: impl IntoRequest<GetConnectionProfileRequest> ) -> Result<Response<ConnectionProfile>, Status>

Use this method to get details about a connection profile.

pub async fn create_connection_profile( &mut self, request: impl IntoRequest<CreateConnectionProfileRequest> ) -> Result<Response<Operation>, Status>

Use this method to create a connection profile in a project and location.

pub async fn update_connection_profile( &mut self, request: impl IntoRequest<UpdateConnectionProfileRequest> ) -> Result<Response<Operation>, Status>

Use this method to update the parameters of a connection profile.

pub async fn delete_connection_profile( &mut self, request: impl IntoRequest<DeleteConnectionProfileRequest> ) -> Result<Response<Operation>, Status>

Use this method to delete a connection profile..

pub async fn discover_connection_profile( &mut self, request: impl IntoRequest<DiscoverConnectionProfileRequest> ) -> Result<Response<DiscoverConnectionProfileResponse>, Status>

Use this method to discover a connection profile. The discover API call exposes the data objects and metadata belonging to the profile. Typically, a request returns children data objects under a parent data object that’s optionally supplied in the request.

pub async fn list_streams( &mut self, request: impl IntoRequest<ListStreamsRequest> ) -> Result<Response<ListStreamsResponse>, Status>

Use this method to list streams in a project and location.

pub async fn get_stream( &mut self, request: impl IntoRequest<GetStreamRequest> ) -> Result<Response<Stream>, Status>

Use this method to get details about a stream.

pub async fn create_stream( &mut self, request: impl IntoRequest<CreateStreamRequest> ) -> Result<Response<Operation>, Status>

Use this method to create a stream.

pub async fn update_stream( &mut self, request: impl IntoRequest<UpdateStreamRequest> ) -> Result<Response<Operation>, Status>

Use this method to update the configuration of a stream.

pub async fn delete_stream( &mut self, request: impl IntoRequest<DeleteStreamRequest> ) -> Result<Response<Operation>, Status>

Use this method to delete a stream.

pub async fn fetch_errors( &mut self, request: impl IntoRequest<FetchErrorsRequest> ) -> Result<Response<Operation>, Status>

Use this method to fetch any errors associated with a stream.

pub async fn fetch_static_ips( &mut self, request: impl IntoRequest<FetchStaticIpsRequest> ) -> Result<Response<FetchStaticIpsResponse>, Status>

The FetchStaticIps API call exposes the static ips used by Datastream. Typically, a request returns children data objects under a parent data object that’s optionally supplied in the request.

pub async fn create_private_connection( &mut self, request: impl IntoRequest<CreatePrivateConnectionRequest> ) -> Result<Response<Operation>, Status>

Use this method to create a private connectivity configuration.

pub async fn get_private_connection( &mut self, request: impl IntoRequest<GetPrivateConnectionRequest> ) -> Result<Response<PrivateConnection>, Status>

Use this method to get details about a private connectivity configuration.

pub async fn list_private_connections( &mut self, request: impl IntoRequest<ListPrivateConnectionsRequest> ) -> Result<Response<ListPrivateConnectionsResponse>, Status>

Use this method to list private connectivity configurations in a project and location.

pub async fn delete_private_connection( &mut self, request: impl IntoRequest<DeletePrivateConnectionRequest> ) -> Result<Response<Operation>, Status>

Use this method to delete a private connectivity configuration.

pub async fn create_route( &mut self, request: impl IntoRequest<CreateRouteRequest> ) -> Result<Response<Operation>, Status>

Use this method to create a route for a private connectivity in a project and location.

pub async fn get_route( &mut self, request: impl IntoRequest<GetRouteRequest> ) -> Result<Response<Route>, Status>

Use this method to get details about a route.

pub async fn list_routes( &mut self, request: impl IntoRequest<ListRoutesRequest> ) -> Result<Response<ListRoutesResponse>, Status>

Use this method to list routes created for a private connectivity in a project and location.

pub async fn delete_route( &mut self, request: impl IntoRequest<DeleteRouteRequest> ) -> Result<Response<Operation>, Status>

Use this method to delete a route.