Module google_api_proto::google::cloud::deploy::v1

Modules

canary

Nested message and enum types in Canary.

cloud_deploy_client

Generated client implementations.

custom_canary_deployment

Nested message and enum types in CustomCanaryDeployment.

delivery_pipeline

Nested message and enum types in DeliveryPipeline.

deploy_job_run

Nested message and enum types in DeployJobRun.

execution_config

Nested message and enum types in ExecutionConfig.

job

Nested message and enum types in Job.

job_run

Nested message and enum types in JobRun.

kubernetes_config

Nested message and enum types in KubernetesConfig.

phase

Nested message and enum types in Phase.

release

Nested message and enum types in Release.

rollout

Nested message and enum types in Rollout.

runtime_config

Nested message and enum types in RuntimeConfig.

strategy

Nested message and enum types in Strategy.

target

Nested message and enum types in Target.

target_artifact

Nested message and enum types in TargetArtifact.

verify_job_run

Nested message and enum types in VerifyJobRun.

Structs

AbandonReleaseRequest

The request object used by AbandonRelease.

AbandonReleaseResponse

The response object for AbandonRelease.

AdvanceChildRolloutJob

An advanceChildRollout Job.

AdvanceChildRolloutJobRun

AdvanceChildRolloutJobRun contains information specific to a advanceChildRollout JobRun.

AdvanceRolloutRequest

The request object used by AdvanceRollout.

AdvanceRolloutResponse

The response object from AdvanceRollout.

AnthosCluster

Information specifying an Anthos Cluster.

ApproveRolloutRequest

The request object used by ApproveRollout.

ApproveRolloutResponse

The response object from ApproveRollout.

BuildArtifact

Description of an a image to use during Skaffold rendering.

Canary

Canary represents the canary deployment strategy.

CanaryDeployment

CanaryDeployment represents the canary deployment configuration

CancelRolloutRequest

The request object used by CancelRollout.

CancelRolloutResponse

The response object from CancelRollout.

ChildRolloutJobs

ChildRollouts job composition

CloudRunConfig

CloudRunConfig contains the Cloud Run runtime configuration.

CloudRunLocation

Information specifying where to deploy a Cloud Run Service.

CloudRunMetadata

CloudRunMetadata contains information from a Cloud Run deployment.

CloudRunRenderMetadata

CloudRunRenderMetadata contains Cloud Run information associated with a Release render.

Config

Service-wide configuration.

CreateChildRolloutJob

A createChildRollout Job.

CreateChildRolloutJobRun

CreateChildRolloutJobRun contains information specific to a createChildRollout JobRun.

CreateDeliveryPipelineRequest

The request object for CreateDeliveryPipeline.

CreateReleaseRequest

The request object for CreateRelease,

CreateRolloutRequest

CreateRolloutRequest is the request object used by CreateRollout.

CreateTargetRequest

The request object for CreateTarget.

CustomCanaryDeployment

CustomCanaryDeployment represents the custom canary deployment configuration.

DefaultPool

Execution using the default Cloud Build pool.

DeleteDeliveryPipelineRequest

The request object for DeleteDeliveryPipeline.

DeleteTargetRequest

The request object for DeleteTarget.

DeliveryPipeline

A DeliveryPipeline resource in the Google Cloud Deploy API.

DeliveryPipelineNotificationEvent

Payload proto for “clouddeploy.googleapis.com/deliverypipeline_notification” Platform Log event that describes the failure to send delivery pipeline status change Pub/Sub notification.

DeployArtifact

The artifacts produced by a deploy operation.

DeployJob

A deploy Job.

DeployJobRun

DeployJobRun contains information specific to a deploy JobRun.

DeployJobRunMetadata

DeployJobRunMetadata surfaces information associated with a DeployJobRun to the user.

DeployParameters

DeployParameters contains deploy parameters information.

DeploymentJobs

Deployment job composition.

ExecutionConfig

Configuration of the environment to use when calling Skaffold.

GetConfigRequest

Request to get a configuration.

GetDeliveryPipelineRequest

The request object for GetDeliveryPipeline

GetJobRunRequest

GetJobRunRequest is the request object used by GetJobRun.

GetReleaseRequest

The request object for GetRelease.

GetRolloutRequest

GetRolloutRequest is the request object used by GetRollout.

GetTargetRequest

The request object for GetTarget.

GkeCluster

Information specifying a GKE Cluster.

IgnoreJobRequest

The request object used by IgnoreJob.

IgnoreJobResponse

The response object from IgnoreJob.

Job

Job represents an operation for a Rollout.

JobRun

A JobRun resource in the Google Cloud Deploy API.

JobRunNotificationEvent

Payload proto for “clouddeploy.googleapis.com/jobrun_notification” Platform Log event that describes the failure to send JobRun resource update Pub/Sub notification.

KubernetesConfig

KubernetesConfig contains the Kubernetes runtime configuration.

ListDeliveryPipelinesRequest

The request object for ListDeliveryPipelines.

ListDeliveryPipelinesResponse

The response object from ListDeliveryPipelines.

ListJobRunsRequest

ListJobRunsRequest is the request object used by ListJobRuns.

ListJobRunsResponse

ListJobRunsResponse is the response object returned by ListJobRuns.

ListReleasesRequest

The request object for ListReleases.

ListReleasesResponse

The response object from ListReleases.

ListRolloutsRequest

ListRolloutsRequest is the request object used by ListRollouts.

ListRolloutsResponse

ListRolloutsResponse is the response object reutrned by ListRollouts.

ListTargetsRequest

The request object for ListTargets.

ListTargetsResponse

The response object from ListTargets.

Metadata

Metadata includes information associated with a Rollout.

MultiTarget

Information specifying a multiTarget.

OperationMetadata

Represents the metadata of the long-running operation.

Phase

Phase represents a collection of jobs that are logically grouped together for a Rollout.

PipelineCondition

PipelineCondition contains all conditions relevant to a Delivery Pipeline.

PipelineReadyCondition

PipelineReadyCondition contains information around the status of the Pipeline.

PrivatePool

Execution using a private Cloud Build pool.

Release

A Release resource in the Google Cloud Deploy API.

ReleaseNotificationEvent

Payload proto for “clouddeploy.googleapis.com/release_notification” Platform Log event that describes the failure to send release status change Pub/Sub notification.

ReleaseRenderEvent

Payload proto for “clouddeploy.googleapis.com/release_render” Platform Log event that describes the render status change.

RenderMetadata

RenderMetadata includes information associated with a Release render.

RetryJobRequest

RetryJobRequest is the request object used by RetryJob.

RetryJobResponse

The response object from ‘RetryJob’.

Rollout

A Rollout resource in the Google Cloud Deploy API.

RolloutNotificationEvent

Payload proto for “clouddeploy.googleapis.com/rollout_notification” Platform Log event that describes the failure to send rollout status change Pub/Sub notification.

RuntimeConfig

RuntimeConfig contains the runtime specific configurations for a deployment strategy.

SerialPipeline

SerialPipeline defines a sequential set of stages for a DeliveryPipeline.

SkaffoldVersion

Details of a supported Skaffold version.

Stage

Stage specifies a location to which to deploy.

Standard

Standard represents the standard deployment strategy.

Strategy

Strategy contains deployment strategy information.

Target

A Target resource in the Google Cloud Deploy API.

TargetArtifact

The artifacts produced by a target render operation.

TargetNotificationEvent

Payload proto for “clouddeploy.googleapis.com/target_notification” Platform Log event that describes the failure to send target status change Pub/Sub notification.

TargetsPresentCondition

TargetsPresentCondition contains information on any Targets defined in the Delivery Pipeline that do not actually exist.

TargetsTypeCondition

TargetsTypeCondition contains information on whether the Targets defined in the Delivery Pipeline are of the same type.

TerminateJobRunRequest

The request object used by TerminateJobRun.

TerminateJobRunResponse

The response object from TerminateJobRun.

UpdateDeliveryPipelineRequest

The request object for UpdateDeliveryPipeline.

UpdateTargetRequest

The request object for UpdateTarget.

VerifyJob

A verify Job.

VerifyJobRun

VerifyJobRun contains information specific to a verify JobRun.

Enums

SkaffoldSupportState

The support state of a specific Skaffold version.

Type

Type indicates the type of the log entry and can be used as a filter.