Module google_api_proto::google::cloud::deploy::v1

Modules

automation_rule

Nested message and enum types in AutomationRule.

automation_run

Nested message and enum types in AutomationRun.

canary

Nested message and enum types in Canary.

cloud_deploy_client

Generated client implementations.

custom_canary_deployment

Nested message and enum types in CustomCanaryDeployment.

custom_target_type

Nested message and enum types in CustomTargetType.

delivery_pipeline

Nested message and enum types in DeliveryPipeline.

deploy_job_run

Nested message and enum types in DeployJobRun.

execution_config

Nested message and enum types in ExecutionConfig.

job

Nested message and enum types in Job.

job_run

Nested message and enum types in JobRun.

kubernetes_config

Nested message and enum types in KubernetesConfig.

phase

Nested message and enum types in Phase.

postdeploy_job_run

Nested message and enum types in PostdeployJobRun.

predeploy_job_run

Nested message and enum types in PredeployJobRun.

release

Nested message and enum types in Release.

repair_mode

Nested message and enum types in RepairMode.

repair_phase

Nested message and enum types in RepairPhase.

rollout

Nested message and enum types in Rollout.

rollout_update_event

Nested message and enum types in RolloutUpdateEvent.

runtime_config

Nested message and enum types in RuntimeConfig.

skaffold_modules

Nested message and enum types in SkaffoldModules.

strategy

Nested message and enum types in Strategy.

target

Nested message and enum types in Target.

target_artifact

Nested message and enum types in TargetArtifact.

verify_job_run

Nested message and enum types in VerifyJobRun.

Structs

AbandonReleaseRequest

The request object used by AbandonRelease.

AbandonReleaseResponse

The response object for AbandonRelease.

AdvanceChildRolloutJob

An advanceChildRollout Job.

AdvanceChildRolloutJobRun

AdvanceChildRolloutJobRun contains information specific to a advanceChildRollout JobRun.

AdvanceRolloutOperation

Contains the information of an automated advance-rollout operation.

AdvanceRolloutRequest

The request object used by AdvanceRollout.

AdvanceRolloutResponse

The response object from AdvanceRollout.

AdvanceRolloutRule

The AdvanceRollout automation rule will automatically advance a successful Rollout to the next phase.

AnthosCluster

Information specifying an Anthos Cluster.

ApproveRolloutRequest

The request object used by ApproveRollout.

ApproveRolloutResponse

The response object from ApproveRollout.

Automation

An Automation resource in the Cloud Deploy API.

AutomationEvent

Payload proto for “clouddeploy.googleapis.com/automation” Platform Log event that describes the Automation related events.

AutomationResourceSelector

AutomationResourceSelector contains the information to select the resources to which an Automation is going to be applied.

AutomationRolloutMetadata

AutomationRolloutMetadata contains Automation-related actions that were performed on a rollout.

AutomationRule

AutomationRule defines the automation activities.

AutomationRuleCondition

AutomationRuleCondition contains conditions relevant to an Automation rule.

AutomationRun

An AutomationRun resource in the Cloud Deploy API.

AutomationRunEvent

Payload proto for “clouddeploy.googleapis.com/automation_run” Platform Log event that describes the AutomationRun related events.

BuildArtifact

Description of an a image to use during Skaffold rendering.

Canary

Canary represents the canary deployment strategy.

CanaryDeployment

CanaryDeployment represents the canary deployment configuration

CancelAutomationRunRequest

The request object used by CancelAutomationRun.

CancelAutomationRunResponse

The response object from CancelAutomationRun.

CancelRolloutRequest

The request object used by CancelRollout.

CancelRolloutResponse

The response object from CancelRollout.

ChildRolloutJobs

ChildRollouts job composition

CloudRunConfig

CloudRunConfig contains the Cloud Run runtime configuration.

CloudRunLocation

Information specifying where to deploy a Cloud Run Service.

CloudRunMetadata

CloudRunMetadata contains information from a Cloud Run deployment.

CloudRunRenderMetadata

CloudRunRenderMetadata contains Cloud Run information associated with a Release render.

Config

Service-wide configuration.

CreateAutomationRequest

The request object for CreateAutomation.

CreateChildRolloutJob

A createChildRollout Job.

CreateChildRolloutJobRun

CreateChildRolloutJobRun contains information specific to a createChildRollout JobRun.

CreateCustomTargetTypeRequest

The request object for CreateCustomTargetType.

CreateDeliveryPipelineRequest

The request object for CreateDeliveryPipeline.

CreateReleaseRequest

The request object for CreateRelease,

CreateRolloutRequest

CreateRolloutRequest is the request object used by CreateRollout.

CreateTargetRequest

The request object for CreateTarget.

CustomCanaryDeployment

CustomCanaryDeployment represents the custom canary deployment configuration.

CustomMetadata

CustomMetadata contains information from a user-defined operation.

CustomTarget

Information specifying a Custom Target.

CustomTargetDeployMetadata

CustomTargetDeployMetadata contains information from a Custom Target deploy operation.

CustomTargetSkaffoldActions

CustomTargetSkaffoldActions represents the CustomTargetType configuration using Skaffold custom actions.

CustomTargetType

A CustomTargetType resource in the Cloud Deploy API.

DefaultPool

Execution using the default Cloud Build pool.

DeleteAutomationRequest

The request object for DeleteAutomation.

DeleteCustomTargetTypeRequest

The request object for DeleteCustomTargetType.

DeleteDeliveryPipelineRequest

The request object for DeleteDeliveryPipeline.

DeleteTargetRequest

The request object for DeleteTarget.

DeliveryPipeline

A DeliveryPipeline resource in the Cloud Deploy API.

DeliveryPipelineNotificationEvent

Payload proto for “clouddeploy.googleapis.com/deliverypipeline_notification” Platform Log event that describes the failure to send delivery pipeline status change Pub/Sub notification.

DeployArtifact

The artifacts produced by a deploy operation.

DeployJob

A deploy Job.

DeployJobRun

DeployJobRun contains information specific to a deploy JobRun.

DeployJobRunMetadata

DeployJobRunMetadata surfaces information associated with a DeployJobRun to the user.

DeployParameters

DeployParameters contains deploy parameters information.

DeploymentJobs

Deployment job composition.

ExecutionConfig

Configuration of the environment to use when calling Skaffold.

GetAutomationRequest

The request object for GetAutomation

GetAutomationRunRequest

The request object for GetAutomationRun

GetConfigRequest

Request to get a configuration.

GetCustomTargetTypeRequest

The request object for GetCustomTargetType.

GetDeliveryPipelineRequest

The request object for GetDeliveryPipeline

GetJobRunRequest

GetJobRunRequest is the request object used by GetJobRun.

GetReleaseRequest

The request object for GetRelease.

GetRolloutRequest

GetRolloutRequest is the request object used by GetRollout.

GetTargetRequest

The request object for GetTarget.

GkeCluster

Information specifying a GKE Cluster.

IgnoreJobRequest

The request object used by IgnoreJob.

IgnoreJobResponse

The response object from IgnoreJob.

Job

Job represents an operation for a Rollout.

JobRun

A JobRun resource in the Cloud Deploy API.

JobRunNotificationEvent

Payload proto for “clouddeploy.googleapis.com/jobrun_notification” Platform Log event that describes the failure to send JobRun resource update Pub/Sub notification.

KubernetesConfig

KubernetesConfig contains the Kubernetes runtime configuration.

ListAutomationRunsRequest

The request object for ListAutomationRuns.

ListAutomationRunsResponse

The response object from ListAutomationRuns.

ListAutomationsRequest

The request object for ListAutomations.

ListAutomationsResponse

The response object from ListAutomations.

ListCustomTargetTypesRequest

The request object for ListCustomTargetTypes.

ListCustomTargetTypesResponse

The response object from ListCustomTargetTypes.

ListDeliveryPipelinesRequest

The request object for ListDeliveryPipelines.

ListDeliveryPipelinesResponse

The response object from ListDeliveryPipelines.

ListJobRunsRequest

ListJobRunsRequest is the request object used by ListJobRuns.

ListJobRunsResponse

ListJobRunsResponse is the response object returned by ListJobRuns.

ListReleasesRequest

The request object for ListReleases.

ListReleasesResponse

The response object from ListReleases.

ListRolloutsRequest

ListRolloutsRequest is the request object used by ListRollouts.

ListRolloutsResponse

ListRolloutsResponse is the response object reutrned by ListRollouts.

ListTargetsRequest

The request object for ListTargets.

ListTargetsResponse

The response object from ListTargets.

Metadata

Metadata includes information associated with a Rollout.

MultiTarget

Information specifying a multiTarget.

OperationMetadata

Represents the metadata of the long-running operation.

Phase

Phase represents a collection of jobs that are logically grouped together for a Rollout.

PipelineCondition

PipelineCondition contains all conditions relevant to a Delivery Pipeline.

PipelineReadyCondition

PipelineReadyCondition contains information around the status of the Pipeline.

Postdeploy

Postdeploy contains the postdeploy job configuration information.

PostdeployJob

A postdeploy Job.

PostdeployJobRun

PostdeployJobRun contains information specific to a postdeploy JobRun.

Predeploy

Predeploy contains the predeploy job configuration information.

PredeployJob

A predeploy Job.

PredeployJobRun

PredeployJobRun contains information specific to a predeploy JobRun.

PrivatePool

Execution using a private Cloud Build pool.

PromoteReleaseOperation

Contains the information of an automated promote-release operation.

PromoteReleaseRule

PromoteRelease rule will automatically promote a release from the current target to a specified target.

Release

A Release resource in the Cloud Deploy API.

ReleaseNotificationEvent

Payload proto for “clouddeploy.googleapis.com/release_notification” Platform Log event that describes the failure to send release status change Pub/Sub notification.

ReleaseRenderEvent

Payload proto for “clouddeploy.googleapis.com/release_render” Platform Log event that describes the render status change.

RenderMetadata

RenderMetadata includes information associated with a Release render.

RepairMode

Configuration of the repair action.

RepairPhase

RepairPhase tracks the repair attempts that have been made for each RepairMode specified in the Automation resource.

RepairRolloutOperation

Contains the information for an automated repair rollout operation.

RepairRolloutRule

The RepairRolloutRule automation rule will automatically repair a failed Rollout.

Retry

Retries the failed job.

RetryAttempt

RetryAttempt represents an action of retrying the failed Cloud Deploy job.

RetryJobRequest

RetryJobRequest is the request object used by RetryJob.

RetryJobResponse

The response object from ‘RetryJob’.

RetryPhase

RetryPhase contains the retry attempts and the metadata for initiating a new attempt.

Rollback

Rolls back a Rollout.

RollbackAttempt

RollbackAttempt represents an action of rolling back a Cloud Deploy ‘Target’.

RollbackTargetConfig

Configs for the Rollback rollout.

RollbackTargetRequest

The request object for RollbackTarget.

RollbackTargetResponse

The response object from RollbackTarget.

Rollout

A Rollout resource in the Cloud Deploy API.

RolloutNotificationEvent

Payload proto for “clouddeploy.googleapis.com/rollout_notification” Platform Log event that describes the failure to send rollout status change Pub/Sub notification.

RolloutUpdateEvent

Payload proto for “clouddeploy.googleapis.com/rollout_update” Platform Log event that describes the rollout update event.

RuntimeConfig

RuntimeConfig contains the runtime specific configurations for a deployment strategy.

SerialPipeline

SerialPipeline defines a sequential set of stages for a DeliveryPipeline.

SkaffoldModules

Skaffold Config modules and their remote source.

SkaffoldVersion

Details of a supported Skaffold version.

Stage

Stage specifies a location to which to deploy.

Standard

Standard represents the standard deployment strategy.

Strategy

Strategy contains deployment strategy information.

Target

A Target resource in the Cloud Deploy API.

TargetArtifact

The artifacts produced by a target render operation.

TargetAttribute

Contains criteria for selecting Targets.

TargetNotificationEvent

Payload proto for “clouddeploy.googleapis.com/target_notification” Platform Log event that describes the failure to send target status change Pub/Sub notification.

TargetsPresentCondition

TargetsPresentCondition contains information on any Targets referenced in the Delivery Pipeline that do not actually exist.

TargetsTypeCondition

TargetsTypeCondition contains information on whether the Targets defined in the Delivery Pipeline are of the same type.

TerminateJobRunRequest

The request object used by TerminateJobRun.

TerminateJobRunResponse

The response object from TerminateJobRun.

UpdateAutomationRequest

The request object for UpdateAutomation.

UpdateCustomTargetTypeRequest

The request object for UpdateCustomTargetType.

UpdateDeliveryPipelineRequest

The request object for UpdateDeliveryPipeline.

UpdateTargetRequest

The request object for UpdateTarget.

VerifyJob

A verify Job.

VerifyJobRun

VerifyJobRun contains information specific to a verify JobRun.

Enums

BackoffMode

The pattern of how wait time is increased.

RepairState

Valid state of a repair attempt.

SkaffoldSupportState

The support state of a specific Skaffold version.

Type

Type indicates the type of the log entry and can be used as a filter.