Module google_api_proto::google::cloud::dialogflow::v2::intent::message

Nested message and enum types in Message.

Modules

basic_card

Nested message and enum types in BasicCard.

browse_carousel_card

Nested message and enum types in BrowseCarouselCard.

card

Nested message and enum types in Card.

carousel_select

Nested message and enum types in CarouselSelect.

column_properties

Nested message and enum types in ColumnProperties.

list_select

Nested message and enum types in ListSelect.

media_content

Nested message and enum types in MediaContent.

Structs

BasicCard

The basic card message. Useful for displaying information.

BrowseCarouselCard

Browse Carousel Card for Actions on Google. https://developers.google.com/actions/assistant/responses#browsing_carousel

Card

The card response message.

CarouselSelect

The card for presenting a carousel of options to select from.

ColumnProperties

Column properties for [TableCard][google.cloud.dialogflow.v2.Intent.Message.TableCard].

Image

The image response message.

LinkOutSuggestion

The suggestion chip message that allows the user to jump out to the app or website associated with this agent.

ListSelect

The card for presenting a list of options to select from.

MediaContent

The media content card for Actions on Google.

QuickReplies

The quick replies response message.

SelectItemInfo

Additional info about the select item for when it is triggered in a dialog.

SimpleResponse

The simple response message containing speech or text.

SimpleResponses

The collection of simple response candidates. This message in QueryResult.fulfillment_messages and WebhookResponse.fulfillment_messages should contain only one SimpleResponse.

Suggestion

The suggestion chip message that the user can tap to quickly post a reply to the conversation.

Suggestions

The collection of suggestions.

TableCard

Table card for Actions on Google.

TableCardCell

Cell of [TableCardRow][google.cloud.dialogflow.v2.Intent.Message.TableCardRow].

TableCardRow

Row of [TableCard][google.cloud.dialogflow.v2.Intent.Message.TableCard].

Text

The text response message.

Enums

Message

Required. The rich response message.

Platform

The rich response message integration platform. See [Integrations](https://cloud.google.com/dialogflow/docs/integrations).