Module google_api_proto::google::cloud::discoveryengine::v1

Modules

big_query_source

Nested message and enum types in BigQuerySource.

complete_query_response

Nested message and enum types in CompleteQueryResponse.

completion_service_client

Generated client implementations.

conversation

Nested message and enum types in Conversation.

conversation_message

Nested message and enum types in ConversationMessage.

conversational_search_service_client

Generated client implementations.

document

Nested message and enum types in Document.

document_info

Nested message and enum types in DocumentInfo.

document_service_client

Generated client implementations.

import_documents_request

Nested message and enum types in ImportDocumentsRequest.

import_error_config

Nested message and enum types in ImportErrorConfig.

import_user_events_request

Nested message and enum types in ImportUserEventsRequest.

interval

Nested message and enum types in Interval.

schema

Nested message and enum types in Schema.

schema_service_client

Generated client implementations.

search_request

Nested message and enum types in SearchRequest.

search_response

Nested message and enum types in SearchResponse.

search_service_client

Generated client implementations.

user_event_service_client

Generated client implementations.

Structs

BigQuerySource

BigQuery source import data from.

CollectUserEventRequest

Request message for CollectUserEvent method.

CompleteQueryRequest

Request message for [CompletionService.CompleteQuery][google.cloud.discoveryengine.v1.CompletionService.CompleteQuery] method.

CompleteQueryResponse

Response message for [CompletionService.CompleteQuery][google.cloud.discoveryengine.v1.CompletionService.CompleteQuery] method.

CompletionInfo

Detailed completion information including completion attribution token and clicked completion info.

Conversation

External conversation proto definition.

ConversationContext

Defines context of the conversation

ConversationMessage

Defines a conversation message.

ConverseConversationRequest

Request message for [ConversationalSearchService.ConverseConversation][google.cloud.discoveryengine.v1.ConversationalSearchService.ConverseConversation] method.

ConverseConversationResponse

Response message for [ConversationalSearchService.ConverseConversation][google.cloud.discoveryengine.v1.ConversationalSearchService.ConverseConversation] method.

CreateConversationRequest

Request for CreateConversation method.

CreateDocumentRequest

Request message for [DocumentService.CreateDocument][google.cloud.discoveryengine.v1.DocumentService.CreateDocument] method.

CreateSchemaMetadata

Metadata for Create Schema LRO.

CreateSchemaRequest

Request message for [SchemaService.CreateSchema][google.cloud.discoveryengine.v1.SchemaService.CreateSchema] method.

CustomAttribute

A custom attribute that is not explicitly modeled in a resource, e.g. [UserEvent][google.cloud.discoveryengine.v1.UserEvent].

DeleteConversationRequest

Request for DeleteConversation method.

DeleteDocumentRequest

Request message for [DocumentService.DeleteDocument][google.cloud.discoveryengine.v1.DocumentService.DeleteDocument] method.

DeleteSchemaMetadata

Metadata for DeleteSchema LRO.

DeleteSchemaRequest

Request message for [SchemaService.DeleteSchema][google.cloud.discoveryengine.v1.SchemaService.DeleteSchema] method.

Document

Document captures all raw metadata information of items to be recommended or searched.

DocumentInfo

Detailed document information associated with a user event.

GcsSource

Cloud Storage location for input content.

GetConversationRequest

Request for GetConversation method.

GetDocumentRequest

Request message for [DocumentService.GetDocument][google.cloud.discoveryengine.v1.DocumentService.GetDocument] method.

GetSchemaRequest

Request message for [SchemaService.GetSchema][google.cloud.discoveryengine.v1.SchemaService.GetSchema] method.

ImportDocumentsMetadata

Metadata related to the progress of the ImportDocuments operation. This is returned by the google.longrunning.Operation.metadata field.

ImportDocumentsRequest

Request message for Import methods.

ImportDocumentsResponse

Response of the [ImportDocumentsRequest][google.cloud.discoveryengine.v1.ImportDocumentsRequest]. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportErrorConfig

Configuration of destination for Import related errors.

ImportUserEventsMetadata

Metadata related to the progress of the Import operation. This is returned by the google.longrunning.Operation.metadata field.

ImportUserEventsRequest

Request message for the ImportUserEvents request.

ImportUserEventsResponse

Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

Interval

A floating point interval.

ListConversationsRequest

Request for ListConversations method.

ListConversationsResponse

Response for ListConversations method.

ListDocumentsRequest

Request message for [DocumentService.ListDocuments][google.cloud.discoveryengine.v1.DocumentService.ListDocuments] method.

ListDocumentsResponse

Response message for [DocumentService.ListDocuments][google.cloud.discoveryengine.v1.DocumentService.ListDocuments] method.

ListSchemasRequest

Request message for [SchemaService.ListSchemas][google.cloud.discoveryengine.v1.SchemaService.ListSchemas] method.

ListSchemasResponse

Response message for [SchemaService.ListSchemas][google.cloud.discoveryengine.v1.SchemaService.ListSchemas] method.

MediaInfo

Media-specific user event information.

PageInfo

Detailed page information.

PanelInfo

Detailed panel information associated with a user event.

PurgeDocumentsMetadata

Metadata related to the progress of the PurgeDocuments operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeDocumentsRequest

Request message for [DocumentService.PurgeDocuments][google.cloud.discoveryengine.v1.DocumentService.PurgeDocuments] method.

PurgeDocumentsResponse

Response message for [DocumentService.PurgeDocuments][google.cloud.discoveryengine.v1.DocumentService.PurgeDocuments] method. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

Reply

Defines a reply message to user.

Schema

Defines the structure and layout of a type of document data.

SearchInfo

Detailed search information.

SearchRequest

Request message for [SearchService.Search][google.cloud.discoveryengine.v1.SearchService.Search] method.

SearchResponse

Response message for [SearchService.Search][google.cloud.discoveryengine.v1.SearchService.Search] method.

TextInput

Defines text input.

TransactionInfo

A transaction represents the entire purchase transaction.

UpdateConversationRequest

Request for UpdateConversation method.

UpdateDocumentRequest

Request message for [DocumentService.UpdateDocument][google.cloud.discoveryengine.v1.DocumentService.UpdateDocument] method.

UpdateSchemaMetadata

Metadata for UpdateSchema LRO.

UpdateSchemaRequest

Request message for [SchemaService.UpdateSchema][google.cloud.discoveryengine.v1.SchemaService.UpdateSchema] method.

UserEvent

UserEvent captures all metadata information Discovery Engine API needs to know about how end users interact with customers’ website.

UserInfo

Information of an end user.

WriteUserEventRequest

Request message for WriteUserEvent method.