Struct google_api_proto::google::cloud::discoveryengine::v1::UserEvent

pub struct UserEvent {

    pub event_type: String,
    pub user_pseudo_id: String,
    pub event_time: Option<Timestamp>,
    pub user_info: Option<UserInfo>,
    pub direct_user_request: bool,
    pub session_id: String,
    pub page_info: Option<PageInfo>,
    pub attribution_token: String,
    pub filter: String,
    pub documents: Vec<DocumentInfo>,
    pub panel: Option<PanelInfo>,
    pub search_info: Option<SearchInfo>,
    pub completion_info: Option<CompletionInfo>,
    pub transaction_info: Option<TransactionInfo>,
    pub tag_ids: Vec<String>,
    pub promotion_ids: Vec<String>,
    pub attributes: BTreeMap<String, CustomAttribute>,
    pub media_info: Option<MediaInfo>,

}

UserEvent captures all metadata information Discovery Engine API needs to know about how end users interact with customers’ website.

Fields
event_type: String

Required. User event type. Allowed values are:

Generic values:

search: Search for Documents.

view-item: Detailed page view of a Document.

view-item-list: View of a panel or ordered list of Documents.

view-home-page: View of the home page.

view-category-page: View of a category page, e.g. Home > Men > Jeans

Retail-related values:

add-to-cart: Add an item(s) to cart, e.g. in Retail online shopping

purchase: Purchase an item(s)

Media-related values:

media-play: Start/resume watching a video, playing a song, etc.

media-complete: Finished or stopped midway through a video, song, etc.

user_pseudo_id: String

Required. A unique identifier for tracking visitors.

For example, this could be implemented with an HTTP cookie, which should be able to uniquely identify a visitor on a single device. This unique identifier should not change if the visitor log in/out of the website.

Do not set the field to the same fixed ID for different users. This mixes the event history of those users together, which results in degraded model quality.

The field must be a UTF-8 encoded string with a length limit of 128 characters. Otherwise, an INVALID_ARGUMENT error is returned.

The field should not contain PII or user-data. We recommend to use Google Analytics Client ID for this field.

event_time: Option<Timestamp>

Only required for [UserEventService.ImportUserEvents][google.cloud.discoveryengine.v1.UserEventService.ImportUserEvents] method. Timestamp of when the user event happened.

user_info: Option<UserInfo>

Information about the end user.

direct_user_request: bool

Should set to true if the request is made directly from the end user, in which case the [UserEvent.user_info.user_agent][google.cloud.discoveryengine.v1.UserInfo.user_agent] can be populated from the HTTP request.

This flag should be set only if the API request is made directly from the end user such as a mobile app (and not if a gateway or a server is processing and pushing the user events).

This should not be set when using the JavaScript tag in [UserEventService.CollectUserEvent][google.cloud.discoveryengine.v1.UserEventService.CollectUserEvent].

session_id: String

A unique identifier for tracking a visitor session with a length limit of 128 bytes. A session is an aggregation of an end user behavior in a time span.

A general guideline to populate the session_id:

If user has no activity for 30 min, a new session_id should be assigned.

The session_id should be unique across users, suggest use uuid or add [UserEvent.user_pseudo_id][google.cloud.discoveryengine.v1.UserEvent.user_pseudo_id] as prefix.

page_info: Option<PageInfo>

Page metadata such as categories and other critical information for certain event types such as view-category-page.

attribution_token: String

Token to attribute an API response to user action(s) to trigger the event.

Highly recommended for user events that are the result of [RecommendationService.Recommend][]. This field enables accurate attribution of recommendation model performance.

The value must be one of:

[RecommendResponse.attribution_token][] for events that are the result of [RecommendationService.Recommend][].

[SearchResponse.attribution_token][google.cloud.discoveryengine.v1.SearchResponse.attribution_token] for events that are the result of [SearchService.Search][google.cloud.discoveryengine.v1.SearchService.Search].

This token enables us to accurately attribute page view or conversion completion back to the event and the particular predict response containing this clicked/purchased product. If user clicks on product K in the recommendation results, pass [RecommendResponse.attribution_token][] as a URL parameter to product K’s page. When recording events on product K’s page, log the [RecommendResponse.attribution_token][] to this field.

filter: String

The filter syntax consists of an expression language for constructing a predicate from one or more fields of the documents being filtered.

One example is for search events, the associated [SearchRequest][google.cloud.discoveryengine.v1.SearchRequest] may contain a filter expression in [SearchRequest.filter][google.cloud.discoveryengine.v1.SearchRequest.filter] conforming to https://google.aip.dev/160#filtering.

Similarly, for view-item-list events that are generated from a [RecommendationService.RecommendRequest][], this field may be populated directly from [RecommendationService.RecommendRequest.filter][] conforming to https://google.aip.dev/160#filtering.

The value must be a UTF-8 encoded string with a length limit of 1,000 characters. Otherwise, an INVALID_ARGUMENT error is returned.

documents: Vec<DocumentInfo>

List of [Document][google.cloud.discoveryengine.v1.Document]s associated with this user event.

This field is optional except for the following event types:

view-item

add-to-cart

purchase

media-play

media-complete

In a search event, this field represents the documents returned to the end user on the current page (the end user may have not finished browsing the whole page yet). When a new page is returned to the end user, after pagination/filtering/ordering even for the same query, a new search event with different [UserEvent.documents][google.cloud.discoveryengine.v1.UserEvent.documents] is desired.

panel: Option<PanelInfo>

Panel metadata associated with this user event.

search_info: Option<SearchInfo>

[SearchService.Search][google.cloud.discoveryengine.v1.SearchService.Search] details related to the event.

This field should be set for search event.

completion_info: Option<CompletionInfo>

[CompletionService.CompleteQuery][google.cloud.discoveryengine.v1.CompletionService.CompleteQuery] details related to the event.

This field should be set for search event when autocomplete function is enabled and the user clicks a suggestion for search.

transaction_info: Option<TransactionInfo>

The transaction metadata (if any) associated with this user event.

tag_ids: Vec<String>

A list of identifiers for the independent experiment groups this user event belongs to. This is used to distinguish between user events associated with different experiment setups on the customer end.

promotion_ids: Vec<String>

The promotion IDs if this is an event associated with promotions. Currently, this field is restricted to at most one ID.

attributes: BTreeMap<String, CustomAttribute>

Extra user event features to include in the recommendation model. These attributes must NOT contain data that needs to be parsed or processed further, e.g. JSON or other encodings.

If you provide custom attributes for ingested user events, also include them in the user events that you associate with prediction requests. Custom attribute formatting must be consistent between imported events and events provided with prediction requests. This lets the Discovery Engine API use those custom attributes when training models and serving predictions, which helps improve recommendation quality.

This field needs to pass all below criteria, otherwise an INVALID_ARGUMENT error is returned:

The key must be a UTF-8 encoded string with a length limit of 5,000 characters.

For text attributes, at most 400 values are allowed. Empty values are not allowed. Each value must be a UTF-8 encoded string with a length limit of 256 characters.

For number attributes, at most 400 values are allowed.

For product recommendations, an example of extra user information is traffic_channel, which is how a user arrives at the site. Users can arrive at the site by coming to the site directly, coming through Google search, or in other ways.

media_info: Option<MediaInfo>

Media-specific info.