Module google_api_proto::google::cloud::documentai::v1beta3

Modules

batch_dataset_documents

Nested message and enum types in BatchDatasetDocuments.

batch_delete_documents_metadata

Nested message and enum types in BatchDeleteDocumentsMetadata.

batch_documents_input_config

Nested message and enum types in BatchDocumentsInputConfig.

batch_process_metadata

Nested message and enum types in BatchProcessMetadata.

batch_process_request

Nested message and enum types in BatchProcessRequest.

common_operation_metadata

Nested message and enum types in CommonOperationMetadata.

dataset

Nested message and enum types in Dataset.

document

Nested message and enum types in Document.

document_id

Nested message and enum types in DocumentId.

document_output_config

Nested message and enum types in DocumentOutputConfig.

document_processor_service_client

Generated client implementations.

document_schema

Nested message and enum types in DocumentSchema.

document_service_client

Generated client implementations.

evaluation

Nested message and enum types in Evaluation.

human_review_status

Nested message and enum types in HumanReviewStatus.

import_documents_metadata

Nested message and enum types in ImportDocumentsMetadata.

import_documents_request

Nested message and enum types in ImportDocumentsRequest.

import_processor_version_request

Nested message and enum types in ImportProcessorVersionRequest.

ocr_config

Nested message and enum types in OcrConfig.

process_options

Nested message and enum types in ProcessOptions.

process_request

Nested message and enum types in ProcessRequest.

processor

Nested message and enum types in Processor.

processor_type

Nested message and enum types in ProcessorType.

processor_version

Nested message and enum types in ProcessorVersion.

review_document_operation_metadata

Nested message and enum types in ReviewDocumentOperationMetadata.

review_document_request

Nested message and enum types in ReviewDocumentRequest.

review_document_response

Nested message and enum types in ReviewDocumentResponse.

revision_ref

Nested message and enum types in RevisionRef.

summary_options

Nested message and enum types in SummaryOptions.

train_processor_version_metadata

Nested message and enum types in TrainProcessorVersionMetadata.

train_processor_version_request

Nested message and enum types in TrainProcessorVersionRequest.

Structs

Barcode

Encodes the detailed information of a barcode.

BatchDatasetDocuments

Dataset documents that the batch operation will be applied to.

BatchDeleteDocumentsMetadata

BatchDeleteDocumentsRequest

BatchDeleteDocumentsResponse

Response of the delete documents operation.

BatchDocumentsInputConfig

The common config to specify a set of documents used as input.

BatchProcessMetadata

The long-running operation metadata for [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments].

BatchProcessRequest

Request message for [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments].

BatchProcessResponse

Response message for [BatchProcessDocuments][google.cloud.documentai.v1beta3.DocumentProcessorService.BatchProcessDocuments].

BoundingPoly

A bounding polygon for the detected image annotation.

CommonOperationMetadata

The common metadata for long running operations.

CreateProcessorRequest

Request message for the [CreateProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.CreateProcessor] method. Notice this request is sent to a regionalized backend service. If the [ProcessorType][google.cloud.documentai.v1beta3.ProcessorType] isn’t available in that region, the creation fails.

Dataset

A singleton resource under a [Processor][google.cloud.documentai.v1beta3.Processor] which configures a collection of documents.

DatasetSchema

Dataset Schema.

DeleteProcessorMetadata

The long-running operation metadata for the [DeleteProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.DeleteProcessor] method.

DeleteProcessorRequest

Request message for the [DeleteProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.DeleteProcessor] method.

DeleteProcessorVersionMetadata

The long-running operation metadata for the [DeleteProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.DeleteProcessorVersion] method.

DeleteProcessorVersionRequest

Request message for the [DeleteProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.DeleteProcessorVersion] method.

DeployProcessorVersionMetadata

The long-running operation metadata for the [DeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.DeployProcessorVersion] method.

DeployProcessorVersionRequest

Request message for the [DeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.DeployProcessorVersion] method.

DeployProcessorVersionResponse

Response message for the [DeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.DeployProcessorVersion] method.

DisableProcessorMetadata

The long-running operation metadata for the [DisableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.DisableProcessor] method.

DisableProcessorRequest

Request message for the [DisableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.DisableProcessor] method.

DisableProcessorResponse

Response message for the [DisableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.DisableProcessor] method. Intentionally empty proto for adding fields in future.

Document

Document represents the canonical document resource in Document AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document AI to iterate and optimize for quality.

DocumentId

Document Identifier.

DocumentMetadata

Metadata about a document.

DocumentOutputConfig

Config that controls the output of documents. All documents will be written as a JSON file.

DocumentPageRange

Range of pages present in a document.

DocumentSchema

The schema defines the output of the processed document by a processor.

EnableProcessorMetadata

The long-running operation metadata for the [EnableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.EnableProcessor] method.

EnableProcessorRequest

Request message for the [EnableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.EnableProcessor] method.

EnableProcessorResponse

Response message for the [EnableProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.EnableProcessor] method. Intentionally empty proto for adding fields in future.

EntityTypeMetadata

Metadata about an entity type.

EvaluateProcessorVersionMetadata

Metadata of the [EvaluateProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.EvaluateProcessorVersion] method.

EvaluateProcessorVersionRequest

Evaluates the given [ProcessorVersion][google.cloud.documentai.v1beta3.ProcessorVersion] against the supplied documents.

EvaluateProcessorVersionResponse

Response of the [EvaluateProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.EvaluateProcessorVersion] method.

Evaluation

An evaluation of a ProcessorVersion’s performance.

EvaluationReference

Gives a short summary of an evaluation, and links to the evaluation itself.

FetchProcessorTypesRequest

Request message for the [FetchProcessorTypes][google.cloud.documentai.v1beta3.DocumentProcessorService.FetchProcessorTypes] method. Some processor types may require the project be added to an allowlist.

FetchProcessorTypesResponse

Response message for the [FetchProcessorTypes][google.cloud.documentai.v1beta3.DocumentProcessorService.FetchProcessorTypes] method.

FieldExtractionMetadata

Metadata for how this field value is extracted.

GcsDocument

Specifies a document stored on Cloud Storage.

GcsDocuments

Specifies a set of documents on Cloud Storage.

GcsPrefix

Specifies all documents on Cloud Storage with a common prefix.

GetDatasetSchemaRequest

Request for GetDatasetSchema.

GetDocumentRequest

GetDocumentResponse

GetEvaluationRequest

Retrieves a specific Evaluation.

GetProcessorRequest

Request message for the [GetProcessor][google.cloud.documentai.v1beta3.DocumentProcessorService.GetProcessor] method.

GetProcessorTypeRequest

Request message for the [GetProcessorType][google.cloud.documentai.v1beta3.DocumentProcessorService.GetProcessorType] method.

GetProcessorVersionRequest

Request message for the [GetProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.GetProcessorVersion] method.

HumanReviewStatus

The status of human review on a processed document.

ImportDocumentsMetadata

Metadata of the import document operation.

ImportDocumentsRequest

ImportDocumentsResponse

Response of the import document operation.

ImportProcessorVersionMetadata

The long-running operation metadata for the [ImportProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.ImportProcessorVersion] method.

ImportProcessorVersionRequest

The request message for the [ImportProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.ImportProcessorVersion] method.

ImportProcessorVersionResponse

The response message for the [ImportProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.ImportProcessorVersion] method.

ListDocumentsRequest

ListDocumentsResponse

ListEvaluationsRequest

Retrieves a list of evaluations for a given [ProcessorVersion][google.cloud.documentai.v1beta3.ProcessorVersion].

ListEvaluationsResponse

The response from ListEvaluations.

ListProcessorTypesRequest

Request message for the [ListProcessorTypes][google.cloud.documentai.v1beta3.DocumentProcessorService.ListProcessorTypes] method. Some processor types may require the project be added to an allowlist.

ListProcessorTypesResponse

Response message for the [ListProcessorTypes][google.cloud.documentai.v1beta3.DocumentProcessorService.ListProcessorTypes] method.

ListProcessorVersionsRequest

Request message for list all processor versions belongs to a processor.

ListProcessorVersionsResponse

Response message for the [ListProcessorVersions][google.cloud.documentai.v1beta3.DocumentProcessorService.ListProcessorVersions] method.

ListProcessorsRequest

Request message for list all processors belongs to a project.

ListProcessorsResponse

Response message for the [ListProcessors][google.cloud.documentai.v1beta3.DocumentProcessorService.ListProcessors] method.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

OcrConfig

Config for Document OCR.

ProcessOptions

Options for Process API

ProcessRequest

Request message for the [ProcessDocument][google.cloud.documentai.v1beta3.DocumentProcessorService.ProcessDocument] method.

ProcessResponse

Response message for the [ProcessDocument][google.cloud.documentai.v1beta3.DocumentProcessorService.ProcessDocument] method.

Processor

The first-class citizen for Document AI. Each processor defines how to extract structural information from a document.

ProcessorType

A processor type is responsible for performing a certain document understanding task on a certain type of document.

ProcessorVersion

A processor version is an implementation of a processor. Each processor can have multiple versions, pretrained by Google internally or uptrained by the customer. A processor can only have one default version at a time. Its document-processing behavior is defined by that version.

ProcessorVersionAlias

Contains the alias and the aliased resource name of processor version.

PropertyMetadata

Metadata about a property.

RawDocument

Payload message of raw document content (bytes).

ReviewDocumentOperationMetadata

The long-running operation metadata for the [ReviewDocument][google.cloud.documentai.v1beta3.DocumentProcessorService.ReviewDocument] method.

ReviewDocumentRequest

Request message for the [ReviewDocument][google.cloud.documentai.v1beta3.DocumentProcessorService.ReviewDocument] method.

ReviewDocumentResponse

Response message for the [ReviewDocument][google.cloud.documentai.v1beta3.DocumentProcessorService.ReviewDocument] method.

RevisionRef

The revision reference specifies which revision on the document to read.

SetDefaultProcessorVersionMetadata

The long-running operation metadata for the [SetDefaultProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.SetDefaultProcessorVersion] method.

SetDefaultProcessorVersionRequest

Request message for the [SetDefaultProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.SetDefaultProcessorVersion] method.

SetDefaultProcessorVersionResponse

Response message for the [SetDefaultProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.SetDefaultProcessorVersion] method.

SummaryOptions

Metadata for document summarization.

TrainProcessorVersionMetadata

The metadata that represents a processor version being created.

TrainProcessorVersionRequest

Request message for the [TrainProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.TrainProcessorVersion] method.

TrainProcessorVersionResponse

The response for [TrainProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.TrainProcessorVersion].

UndeployProcessorVersionMetadata

The long-running operation metadata for the [UndeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.UndeployProcessorVersion] method.

UndeployProcessorVersionRequest

Request message for the [UndeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.UndeployProcessorVersion] method.

UndeployProcessorVersionResponse

Response message for the [UndeployProcessorVersion][google.cloud.documentai.v1beta3.DocumentProcessorService.UndeployProcessorVersion] method.

UpdateDatasetOperationMetadata

UpdateDatasetRequest

UpdateDatasetSchemaRequest

Request for UpdateDatasetSchema.

Vertex

A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.

Enums

DatasetSplitType

Documents belonging to a dataset will be split into different groups referred to as splits: train, test.

DocumentLabelingState

Describes the labelling status of a document.