Module google_api_proto::google::cloud::edgecontainer::v1

Modules

edge_container_client

Generated client implementations.

node_pool

Nested message and enum types in NodePool.

vpn_connection

Nested message and enum types in VpnConnection.

Structs

Authorization

RBAC policy that will be applied and managed by GEC.

Cluster

A Google Distributed Cloud Edge Kubernetes cluster.

ClusterNetworking

Cluster-wide networking configuration.

ClusterUser

A user principal for an RBAC policy.

CreateClusterRequest

Creates a cluster.

CreateNodePoolRequest

Creates a node pool.

CreateVpnConnectionRequest

Creates a VPN connection.

DeleteClusterRequest

Deletes a cluster.

DeleteNodePoolRequest

Deletes a node pool.

DeleteVpnConnectionRequest

Deletes a vpn connection.

Fleet

Fleet related configuration.

GenerateAccessTokenRequest

Generates an access token for a cluster.

GenerateAccessTokenResponse

An access token for a cluster.

GetClusterRequest

Gets a cluster.

GetMachineRequest

Gets a machine.

GetNodePoolRequest

Gets a node pool.

GetVpnConnectionRequest

Gets a VPN connection.

ListClustersRequest

Lists clusters in a location.

ListClustersResponse

List of clusters in a location.

ListMachinesRequest

Lists machines in a site.

ListMachinesResponse

List of machines in a site.

ListNodePoolsRequest

Lists node pools in a cluster.

ListNodePoolsResponse

List of node pools in a cluster.

ListVpnConnectionsRequest

Lists VPN connections.

ListVpnConnectionsResponse

List of VPN connections in a location.

LocationMetadata

Metadata for a given [google.cloud.location.Location][google.cloud.location.Location].

Machine

A Google Distributed Cloud Edge machine capable of acting as a Kubernetes node.

MaintenancePolicy

Maintenance policy configuration.

MaintenanceWindow

Maintenance window configuration

NodePool

A set of Kubernetes nodes in a cluster with common configuration and specification.

OperationMetadata

Long-running operation metadata for Edge Container API methods.

Quota

Represents quota for Edge Container resources.

RecurringTimeWindow

Represents an arbitrary window of time that recurs.

TimeWindow

Represents an arbitrary window of time.

UpdateClusterRequest

Updates a cluster.

UpdateNodePoolRequest

Updates a node pool.

VpnConnection

A VPN connection .

ZoneMetadata

A Google Distributed Cloud Edge zone where edge machines are located.

Enums

KmsKeyState

Represents the accessibility state of a customer-managed KMS key used for CMEK integration.