Module google_api_proto::google::cloud::edgenetwork::v1

Modules

diagnose_network_response

Nested message and enum types in DiagnoseNetworkResponse.

edge_network_client

Generated client implementations.

interconnect

Nested message and enum types in Interconnect.

interconnect_diagnostics

Nested message and enum types in InterconnectDiagnostics.

router

Nested message and enum types in Router.

router_status

Nested message and enum types in RouterStatus.

Structs

CreateInterconnectAttachmentRequest

Message for creating a InterconnectAttachment

CreateNetworkRequest

Message for creating a Network

CreateRouterRequest

Message for creating a Router

CreateSubnetRequest

Message for creating a Subnet

DeleteInterconnectAttachmentRequest

Message for deleting a InterconnectAttachment

DeleteNetworkRequest

Message for deleting a Network

DeleteRouterRequest

Message for deleting a Router

DeleteSubnetRequest

Message for deleting a Subnet

DiagnoseInterconnectRequest

Message for requesting the diagnostics of an interconnect within a specific zone.

DiagnoseInterconnectResponse

DiagnoseInterconnectResponse contains the current diagnostics for a specific interconnect.

DiagnoseNetworkRequest

Message for requesting the diagnostics of a network within a specific zone.

DiagnoseNetworkResponse

DiagnoseNetworkResponse contains the current status for a specific network.

DiagnoseRouterRequest

Message for requesting diagnositcs of a router within a specific zone.

DiagnoseRouterResponse

DiagnoseRouterResponse contains the current status for a specific router.

GetInterconnectAttachmentRequest

Message for getting a InterconnectAttachment

GetInterconnectRequest

Message for getting a Interconnect

GetNetworkRequest

Message for getting a Network

GetRouterRequest

Message for getting a Router

GetSubnetRequest

Message for getting a Subnet

GetZoneRequest

Message for getting a Zone

InitializeZoneRequest

Message for initializing a specified zone

InitializeZoneResponse

The response of initializing a zone

Interconnect

Message describing Interconnect object

InterconnectAttachment

Message describing InterconnectAttachment object

InterconnectDiagnostics

Diagnostics information about interconnect, contains detailed and current technical information about Google’s side of the connection.

LinkLayerAddress

LinkLayerAddress contains an IP address and corresponding link-layer address.

ListInterconnectAttachmentsRequest

Message for requesting list of InterconnectAttachments

ListInterconnectAttachmentsResponse

Message for response to listing InterconnectAttachments

ListInterconnectsRequest

Message for requesting list of Interconnects

ListInterconnectsResponse

Message for response to listing Interconnects

ListNetworksRequest

Message for requesting list of Networks

ListNetworksResponse

Message for response to listing Networks

ListRoutersRequest

Message for requesting list of Routers

ListRoutersResponse

Message for response to listing Routers

ListSubnetsRequest

Message for requesting list of Subnets

ListSubnetsResponse

Message for response to listing Subnets

ListZonesRequest

Message for requesting list of Zones

ListZonesResponse

Message for response to listing Zones

Network

Message describing Network object

OperationMetadata

Represents the metadata of the long-running operation.

Router

Message describing Router object

RouterStatus

Describing the current status of a router.

Subnet

Message describing Subnet object

SubnetStatus

SubnetStatus contains detailed and current technical information about this subnet resource.

UpdateRouterRequest

Message for updating a Router

UpdateSubnetRequest

Message for updating a Subnet

Zone

A Google Edge Cloud zone.

Enums

ResourceState

ResourceState describes the state the resource. A normal lifecycle of a new resource being created would be: PENDING -> PROVISIONING -> RUNNING. A normal lifecycle of an existing resource being deleted would be: RUNNING -> DELETING. Any failures during processing will result the resource to be in a SUSPENDED state.