Module google_api_proto::google::cloud::eventarc::v1

Modules

channel

Nested message and enum types in Channel.

destination

Nested message and enum types in Destination.

eventarc_client

Generated client implementations.

transport

Nested message and enum types in Transport.

Structs

Channel

A representation of the Channel resource. A Channel is a resource on which event providers publish their events. The published events are delivered through the transport associated with the channel. Note that a channel is associated with exactly one event provider.

ChannelConnection

A representation of the ChannelConnection resource. A ChannelConnection is a resource which event providers create during the activation process to establish a connection between the provider and the subscriber channel.

CloudRun

Represents a Cloud Run destination.

CreateChannelConnectionRequest

The request message for the CreateChannelConnection method.

CreateChannelRequest

The request message for the CreateChannel method.

CreateTriggerRequest

The request message for the CreateTrigger method.

DeleteChannelConnectionRequest

The request message for the DeleteChannelConnection method.

DeleteChannelRequest

The request message for the DeleteChannel method.

DeleteTriggerRequest

The request message for the DeleteTrigger method.

Destination

Represents a target of an invocation over HTTP.

EventFilter

Filters events based on exact matches on the CloudEvents attributes.

EventType

A representation of the event type resource.

FilteringAttribute

A representation of the FilteringAttribute resource. Filtering attributes are per event type.

GetChannelConnectionRequest

The request message for the GetChannelConnection method.

GetChannelRequest

The request message for the GetChannel method.

GetGoogleChannelConfigRequest

The request message for the GetGoogleChannelConfig method.

GetProviderRequest

The request message for the GetProvider method.

GetTriggerRequest

The request message for the GetTrigger method.

Gke

Represents a GKE destination.

GoogleChannelConfig

A GoogleChannelConfig is a resource that stores the custom settings respected by Eventarc first-party triggers in the matching region. Once configured, first-party event data will be protected using the specified custom managed encryption key instead of Google-managed encryption keys.

ListChannelConnectionsRequest

The request message for the ListChannelConnections method.

ListChannelConnectionsResponse

The response message for the ListChannelConnections method.

ListChannelsRequest

The request message for the ListChannels method.

ListChannelsResponse

The response message for the ListChannels method.

ListProvidersRequest

The request message for the ListProviders method.

ListProvidersResponse

The response message for the ListProviders method.

ListTriggersRequest

The request message for the ListTriggers method.

ListTriggersResponse

The response message for the ListTriggers method.

OperationMetadata

Represents the metadata of the long-running operation.

Provider

A representation of the Provider resource.

Pubsub

Represents a Pub/Sub transport.

StateCondition

A condition that is part of the trigger state computation.

Transport

Represents the transport intermediaries created for the trigger to deliver events.

Trigger

A representation of the trigger resource.

UpdateChannelRequest

The request message for the UpdateChannel method.

UpdateGoogleChannelConfigRequest

The request message for the UpdateGoogleChannelConfig method.

UpdateTriggerRequest

The request message for the UpdateTrigger method.