Module google_api_proto::google::cloud::eventarc::v1

Modules

channel

Nested message and enum types in Channel.

destination

Nested message and enum types in Destination.

eventarc_client

Generated client implementations.

transport

Nested message and enum types in Transport.

Structs

Channel

A representation of the Channel resource. A Channel is a resource on which event providers publish their events. The published events are delivered through the transport associated with the channel. Note that a channel is associated with exactly one event provider.

ChannelConnection

A representation of the ChannelConnection resource. A ChannelConnection is a resource which event providers create during the activation process to establish a connection between the provider and the subscriber channel.

CloudRun

Represents a Cloud Run destination.

CreateChannelConnectionRequest

The request message for the CreateChannelConnection method.

CreateChannelRequest

The request message for the CreateChannel method.

CreateTriggerRequest

The request message for the CreateTrigger method.

DeleteChannelConnectionRequest

The request message for the DeleteChannelConnection method.

DeleteChannelRequest

The request message for the DeleteChannel method.

DeleteTriggerRequest

The request message for the DeleteTrigger method.

Destination

Represents a target of an invocation over HTTP.

EventFilter

Filters events based on exact matches on the CloudEvents attributes.

GetChannelConnectionRequest

The request message for the GetChannelConnection method.

GetChannelRequest

The request message for the GetChannel method.

GetTriggerRequest

The request message for the GetTrigger method.

Gke

Represents a GKE destination.

ListChannelConnectionsRequest

The request message for the ListChannelConnections method.

ListChannelConnectionsResponse

The response message for the ListChannelConnections method.

ListChannelsRequest

The request message for the ListChannels method.

ListChannelsResponse

The response message for the ListChannels method.

ListTriggersRequest

The request message for the ListTriggers method.

ListTriggersResponse

The response message for the ListTriggers method.

OperationMetadata

Represents the metadata of the long-running operation.

Pubsub

Represents a Pub/Sub transport.

Transport

Represents the transport intermediaries created for the trigger to deliver events.

Trigger

A representation of the trigger resource.

UpdateChannelRequest

The request message for the UpdateChannel method.

UpdateTriggerRequest

The request message for the UpdateTrigger method.