Module google_api_proto::google::cloud::gkebackup::v1

Modules

backup

Nested message and enum types in Backup.

backup_for_gke_client

Generated client implementations.

backup_plan

Nested message and enum types in BackupPlan.

exclusion_window

Nested message and enum types in ExclusionWindow.

restore

Nested message and enum types in Restore.

restore_config

Nested message and enum types in RestoreConfig.

restore_plan

Nested message and enum types in RestorePlan.

volume_backup

Nested message and enum types in VolumeBackup.

volume_data_restore_policy_override

Nested message and enum types in VolumeDataRestorePolicyOverride.

volume_restore

Nested message and enum types in VolumeRestore.

volume_type_enum

Nested message and enum types in VolumeTypeEnum.

Structs

Backup

Represents a request to perform a single point-in-time capture of some portion of the state of a GKE cluster, the record of the backup operation itself, and an anchor for the underlying artifacts that comprise the Backup (the config backup and VolumeBackups).

BackupPlan

Defines the configuration and scheduling for a “line” of Backups.

CreateBackupPlanRequest

Request message for CreateBackupPlan.

CreateBackupRequest

Request message for CreateBackup.

CreateRestorePlanRequest

Request message for CreateRestorePlan.

CreateRestoreRequest

Request message for CreateRestore.

DeleteBackupPlanRequest

Request message for DeleteBackupPlan.

DeleteBackupRequest

Request message for DeleteBackup.

DeleteRestorePlanRequest

Request message for DeleteRestorePlan.

DeleteRestoreRequest

Request message for DeleteRestore.

EncryptionKey

Defined a customer managed encryption key that will be used to encrypt Backup artifacts.

ExclusionWindow

Defines a time window during which no backup should happen. All time and date are in UTC.

GetBackupIndexDownloadUrlRequest

Request message for GetBackupIndexDownloadUrl.

GetBackupIndexDownloadUrlResponse

Response message for GetBackupIndexDownloadUrl.

GetBackupPlanRequest

Request message for GetBackupPlan.

GetBackupRequest

Request message for GetBackup.

GetRestorePlanRequest

Request message for GetRestorePlan.

GetRestoreRequest

Request message for GetRestore.

GetVolumeBackupRequest

Request message for GetVolumeBackup.

GetVolumeRestoreRequest

Request message for GetVolumeRestore.

ListBackupPlansRequest

Request message for ListBackupPlans.

ListBackupPlansResponse

Response message for ListBackupPlans.

ListBackupsRequest

Request message for ListBackups.

ListBackupsResponse

Response message for ListBackups.

ListRestorePlansRequest

Request message for ListRestorePlans.

ListRestorePlansResponse

Response message for ListRestorePlans.

ListRestoresRequest

Request message for ListRestores.

ListRestoresResponse

Response message for ListRestores.

ListVolumeBackupsRequest

Request message for ListVolumeBackups.

ListVolumeBackupsResponse

Response message for ListVolumeBackups.

ListVolumeRestoresRequest

Request message for ListVolumeRestores.

ListVolumeRestoresResponse

Response message for ListVolumeRestores.

NamespacedName

A reference to a namespaced resource in Kubernetes.

NamespacedNames

A list of namespaced Kubernetes resources.

Namespaces

A list of Kubernetes Namespaces

OperationMetadata

Represents the metadata of the long-running operation.

ResourceSelector

Defines a selector to identify a single or a group of resources. Conditions in the selector are optional, but at least one field should be set to a non-empty value. If a condition is not specified, no restrictions will be applied on that dimension. If more than one condition is specified, a resource will be selected if and only if all conditions are met.

Restore

Represents both a request to Restore some portion of a Backup into a target GKE cluster and a record of the restore operation itself.

RestoreConfig

Configuration of a restore.

RestorePlan

The configuration of a potential series of Restore operations to be performed against Backups belong to a particular BackupPlan.

RpoConfig

Defines RPO scheduling configuration for automatically creating Backups via this BackupPlan.

UpdateBackupPlanRequest

Request message for UpdateBackupPlan.

UpdateBackupRequest

Request message for UpdateBackup.

UpdateRestorePlanRequest

Request message for UpdateRestorePlan.

UpdateRestoreRequest

Request message for UpdateRestore.

VolumeBackup

Represents the backup of a specific persistent volume as a component of a Backup - both the record of the operation and a pointer to the underlying storage-specific artifacts.

VolumeDataRestorePolicyOverride

Defines an override to apply a VolumeDataRestorePolicy for scoped resources.

VolumeRestore

Represents the operation of restoring a volume from a VolumeBackup.

VolumeTypeEnum

Message to encapsulate VolumeType enum.