Module google_api_proto::google::cloud::gkebackup::v1::restore_config

Nested message and enum types in RestoreConfig.

Structs

ClusterResourceRestoreScope

Identifies the cluster-scoped resources to restore from the Backup.

GroupKind

This is a direct map to the Kubernetes GroupKind type [GroupKind](https://godoc.org/k8s.io/apimachinery/pkg/runtime/schema#GroupKind) and is used for identifying specific “types” of resources to restore.

SubstitutionRule

A transformation rule to be applied against Kubernetes resources as they are selected for restoration from a Backup. A rule contains both filtering logic (which resources are subject to substitution) and substitution logic.

Enums

ClusterResourceConflictPolicy

Defines the behavior for handling the situation where cluster-scoped resources being restored already exist in the target cluster.

NamespacedResourceRestoreMode

Defines the behavior for handling the situation where sets of namespaced resources being restored already exist in the target cluster.

NamespacedResourceRestoreScope

Specifies the namespaced resources to restore from the Backup. Only one of the entries may be specified. If not specified, NO namespaced resources will be restored.

VolumeDataRestorePolicy

Defines how volume data should be restored