Module google_api_proto::google::cloud::gkemulticloud::v1

Modules

attached_cluster

Nested message and enum types in AttachedCluster.

attached_clusters_client

Generated client implementations.

aws_cluster

Nested message and enum types in AwsCluster.

aws_clusters_client

Generated client implementations.

aws_instance_placement

Nested message and enum types in AwsInstancePlacement.

aws_node_pool

Nested message and enum types in AwsNodePool.

aws_volume_template

Nested message and enum types in AwsVolumeTemplate.

azure_cluster

Nested message and enum types in AzureCluster.

azure_clusters_client

Generated client implementations.

azure_node_pool

Nested message and enum types in AzureNodePool.

binary_authorization

Nested message and enum types in BinaryAuthorization.

logging_component_config

Nested message and enum types in LoggingComponentConfig.

node_taint

Nested message and enum types in NodeTaint.

Structs

AttachedCluster

An Anthos cluster running on customer own infrastructure.

AttachedClusterError

AttachedClusterError describes errors found on attached clusters.

AttachedClusterGroup

Identities of a group-type subject for Attached clusters.

AttachedClusterUser

Identities of a user-type subject for Attached clusters.

AttachedClustersAuthorization

Configuration related to the cluster RBAC settings.

AttachedOidcConfig

OIDC discovery information of the target cluster.

AttachedPlatformVersionInfo

Information about a supported Attached Clusters platform version.

AttachedProxyConfig

Details of a proxy config.

AttachedServerConfig

AttachedServerConfig provides information about supported Kubernetes versions

AwsAuthorization

Configuration related to the cluster RBAC settings.

AwsAutoscalingGroupMetricsCollection

Configuration related to CloudWatch metrics collection in an AWS Auto Scaling group.

AwsCluster

An Anthos cluster running on AWS.

AwsClusterError

AwsClusterError describes errors found on AWS clusters.

AwsClusterGroup

Identities of a group-type subject for AWS clusters.

AwsClusterNetworking

ClusterNetworking defines cluster-wide networking configuration.

AwsClusterUser

Identities of a user-type subject for AWS clusters.

AwsConfigEncryption

Config encryption for user data.

AwsControlPlane

ControlPlane defines common parameters between control plane nodes.

AwsDatabaseEncryption

Configuration related to application-layer secrets encryption.

AwsInstancePlacement

Details of placement information for an instance. Limitations for using the host tenancy:

AwsJsonWebKeys

AwsJsonWebKeys is a valid JSON Web Key Set as specififed in RFC 7517.

AwsK8sVersionInfo

Kubernetes version information of GKE cluster on AWS.

AwsNodeConfig

Parameters that describe the nodes in a cluster.

AwsNodeManagement

AwsNodeManagement defines the set of node management features turned on for an AWS node pool.

AwsNodePool

An Anthos node pool running on AWS.

AwsNodePoolAutoscaling

AwsNodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

AwsNodePoolError

AwsNodePoolError describes errors found on AWS node pools.

AwsOpenIdConfig

AwsOpenIdConfig is an OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

AwsProxyConfig

Details of a proxy config stored in AWS Secret Manager.

AwsServerConfig

AwsServerConfig is the configuration of GKE cluster on AWS.

AwsServicesAuthentication

Authentication configuration for the management of AWS resources.

AwsSshConfig

SSH configuration for AWS resources.

AwsVolumeTemplate

Configuration template for AWS EBS volumes.

AzureAuthorization

Configuration related to the cluster RBAC settings.

AzureClient

AzureClient resources hold client authentication information needed by the Anthos Multi-Cloud API to manage Azure resources on your Azure subscription.

AzureCluster

An Anthos cluster running on Azure.

AzureClusterError

AzureClusterError describes errors found on Azure clusters.

AzureClusterGroup

Identities of a group-type subject for Azure clusters.

AzureClusterNetworking

ClusterNetworking contains cluster-wide networking configuration.

AzureClusterResources

Managed Azure resources for the cluster.

AzureClusterUser

Identities of a user-type subject for Azure clusters.

AzureConfigEncryption

Configuration related to config data encryption.

AzureControlPlane

AzureControlPlane represents the control plane configurations.

AzureDatabaseEncryption

Configuration related to application-layer secrets encryption.

AzureDiskTemplate

Configuration for Azure Disks.

AzureJsonWebKeys

AzureJsonWebKeys is a valid JSON Web Key Set as specififed in RFC 7517.

AzureK8sVersionInfo

Kubernetes version information of GKE cluster on Azure.

AzureNodeConfig

Parameters that describe the configuration of all node machines on a given node pool.

AzureNodeManagement

AzureNodeManagement defines the set of node management features turned on for an Azure node pool.

AzureNodePool

An Anthos node pool running on Azure.

AzureNodePoolAutoscaling

Configuration related to Kubernetes cluster autoscaler.

AzureNodePoolError

AzureNodePoolError describes errors found on Azure node pools.

AzureOpenIdConfig

AzureOpenIdConfig is an OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

AzureProxyConfig

Details of a proxy config stored in Azure Key Vault.

AzureServerConfig

AzureServerConfig contains information about a Google Cloud location, such as supported Azure regions and Kubernetes versions.

AzureServicesAuthentication

Authentication configuration for the management of Azure resources.

AzureSshConfig

SSH configuration for Azure resources.

BinaryAuthorization

Configuration for Binary Authorization.

CreateAttachedClusterRequest

Request message for AttachedClusters.CreateAttachedCluster method.

CreateAwsClusterRequest

Request message for AwsClusters.CreateAwsCluster method.

CreateAwsNodePoolRequest

Response message for AwsClusters.CreateAwsNodePool method.

CreateAzureClientRequest

Request message for AzureClusters.CreateAzureClient method.

CreateAzureClusterRequest

Request message for AzureClusters.CreateAzureCluster method.

CreateAzureNodePoolRequest

Response message for AzureClusters.CreateAzureNodePool method.

DeleteAttachedClusterRequest

Request message for AttachedClusters.DeleteAttachedCluster method.

DeleteAwsClusterRequest

Request message for AwsClusters.DeleteAwsCluster method.

DeleteAwsNodePoolRequest

Request message for AwsClusters.DeleteAwsNodePool method.

DeleteAzureClientRequest

Request message for AzureClusters.DeleteAzureClient method.

DeleteAzureClusterRequest

Request message for AzureClusters.DeleteAzureCluster method.

DeleteAzureNodePoolRequest

Request message for AzureClusters.DeleteAzureNodePool method.

Fleet

Fleet related configuration.

GenerateAttachedClusterAgentTokenRequest

GenerateAttachedClusterAgentTokenResponse

GenerateAttachedClusterInstallManifestRequest

Request message for AttachedClusters.GenerateAttachedClusterInstallManifest method.

GenerateAttachedClusterInstallManifestResponse

Response message for AttachedClusters.GenerateAttachedClusterInstallManifest method.

GenerateAwsAccessTokenRequest

Request message for AwsClusters.GenerateAwsAccessToken method.

GenerateAwsAccessTokenResponse

Response message for AwsClusters.GenerateAwsAccessToken method.

GenerateAwsClusterAgentTokenRequest

GenerateAwsClusterAgentTokenResponse

GenerateAzureAccessTokenRequest

Request message for AzureClusters.GenerateAzureAccessToken method.

GenerateAzureAccessTokenResponse

Response message for AzureClusters.GenerateAzureAccessToken method.

GenerateAzureClusterAgentTokenRequest

GenerateAzureClusterAgentTokenResponse

GetAttachedClusterRequest

Request message for AttachedClusters.GetAttachedCluster method.

GetAttachedServerConfigRequest

GetAttachedServerConfigRequest gets the server config for attached clusters.

GetAwsClusterRequest

Request message for AwsClusters.GetAwsCluster method.

GetAwsJsonWebKeysRequest

GetAwsJsonWebKeysRequest gets the public component of the keys used by the cluster to sign token requests. This will be the jwks_uri for the discover document returned by getOpenIDConfig. See the OpenID Connect Discovery 1.0 specification for details.

GetAwsNodePoolRequest

Request message for AwsClusters.GetAwsNodePool method.

GetAwsOpenIdConfigRequest

GetAwsOpenIdConfigRequest gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

GetAwsServerConfigRequest

GetAwsServerConfigRequest gets the server config of GKE cluster on AWS.

GetAzureClientRequest

Request message for AzureClusters.GetAzureClient method.

GetAzureClusterRequest

Request message for AzureClusters.GetAzureCluster method.

GetAzureJsonWebKeysRequest

GetAzureJsonWebKeysRequest gets the public component of the keys used by the cluster to sign token requests. This will be the jwks_uri for the discover document returned by getOpenIDConfig. See the OpenID Connect Discovery 1.0 specification for details.

GetAzureNodePoolRequest

Request message for AzureClusters.GetAzureNodePool method.

GetAzureOpenIdConfigRequest

GetAzureOpenIdConfigRequest gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

GetAzureServerConfigRequest

GetAzureServerConfigRequest gets the server config of GKE cluster on Azure.

ImportAttachedClusterRequest

Request message for AttachedClusters.ImportAttachedCluster method.

Jwk

Jwk is a JSON Web Key as specified in RFC 7517.

KubernetesSecret

Information about a Kubernetes Secret

ListAttachedClustersRequest

Request message for AttachedClusters.ListAttachedClusters method.

ListAttachedClustersResponse

Response message for AttachedClusters.ListAttachedClusters method.

ListAwsClustersRequest

Request message for AwsClusters.ListAwsClusters method.

ListAwsClustersResponse

Response message for AwsClusters.ListAwsClusters method.

ListAwsNodePoolsRequest

Request message for AwsClusters.ListAwsNodePools method.

ListAwsNodePoolsResponse

Response message for AwsClusters.ListAwsNodePools method.

ListAzureClientsRequest

Request message for AzureClusters.ListAzureClients method.

ListAzureClientsResponse

Response message for AzureClusters.ListAzureClients method.

ListAzureClustersRequest

Request message for AzureClusters.ListAzureClusters method.

ListAzureClustersResponse

Response message for AzureClusters.ListAzureClusters method.

ListAzureNodePoolsRequest

Request message for AzureClusters.ListAzureNodePools method.

ListAzureNodePoolsResponse

Response message for AzureClusters.ListAzureNodePools method.

LoggingComponentConfig

Parameters that describe the Logging component configuration in a cluster.

LoggingConfig

Parameters that describe the Logging configuration in a cluster.

ManagedPrometheusConfig

ManagedPrometheusConfig defines the configuration for Google Cloud Managed Service for Prometheus.

MaxPodsConstraint

Constraints applied to pods.

MonitoringConfig

Parameters that describe the Monitoring configuration in a cluster.

NodeTaint

The taint content for the node taint.

OperationMetadata

Metadata about a long-running operation.

ReplicaPlacement

Configuration for the placement of a control plane replica.

RollbackAwsNodePoolUpdateRequest

Request message for AwsClusters.RollbackAwsNodePoolUpdate method.

SpotConfig

SpotConfig has configuration info for Spot node.

SurgeSettings

SurgeSettings contains the parameters for Surge update.

UpdateAttachedClusterRequest

Request message for AttachedClusters.UpdateAttachedCluster method.

UpdateAwsClusterRequest

Request message for AwsClusters.UpdateAwsCluster method.

UpdateAwsNodePoolRequest

Request message for AwsClusters.UpdateAwsNodePool method.

UpdateAzureClusterRequest

Request message for AzureClusters.UpdateAzureCluster method.

UpdateAzureNodePoolRequest

Request message for AzureClusters.UpdateAzureNodePool method.

UpdateSettings

UpdateSettings control the level of parallelism and the level of disruption caused during the update of a node pool.

WorkloadIdentityConfig

Workload Identity settings.