Module google_api_proto::google::cloud::gkemulticloud::v1

Modules

aws_cluster

Nested message and enum types in AwsCluster.

aws_clusters_client

Generated client implementations.

aws_instance_placement

Nested message and enum types in AwsInstancePlacement.

aws_node_pool

Nested message and enum types in AwsNodePool.

aws_volume_template

Nested message and enum types in AwsVolumeTemplate.

azure_cluster

Nested message and enum types in AzureCluster.

azure_clusters_client

Generated client implementations.

azure_node_pool

Nested message and enum types in AzureNodePool.

logging_component_config

Nested message and enum types in LoggingComponentConfig.

node_taint

Nested message and enum types in NodeTaint.

Structs

AwsAuthorization

Configuration related to the cluster RBAC settings.

AwsCluster

An Anthos cluster running on AWS.

AwsClusterNetworking

ClusterNetworking defines cluster-wide networking configuration.

AwsClusterUser

Identities of a user-type subject for AWS clusters.

AwsConfigEncryption

Config encryption for user data.

AwsControlPlane

ControlPlane defines common parameters between control plane nodes.

AwsDatabaseEncryption

Configuration related to application-layer secrets encryption.

AwsInstancePlacement

Details of placement information for an instance. Limitation for using the host tenancy.

AwsK8sVersionInfo

Kubernetes version information of GKE cluster on AWS.

AwsNodeConfig

Parameters that describe the nodes in a cluster.

AwsNodePool

An Anthos node pool running on AWS.

AwsNodePoolAutoscaling

AwsNodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

AwsProxyConfig

Details of a proxy config stored in AWS Secret Manager.

AwsServerConfig

AwsServerConfig is the configuration of GKE cluster on AWS.

AwsServicesAuthentication

Authentication configuration for the management of AWS resources.

AwsSshConfig

SSH configuration for AWS resources.

AwsVolumeTemplate

Configuration template for AWS EBS volumes.

AzureAuthorization

Configuration related to the cluster RBAC settings.

AzureClient

AzureClient resources hold client authentication information needed by the Anthos Multi-Cloud API to manage Azure resources on your Azure subscription.

AzureCluster

An Anthos cluster running on Azure.

AzureClusterNetworking

ClusterNetworking contains cluster-wide networking configuration.

AzureClusterResources

Managed Azure resources for the cluster.

AzureClusterUser

Identities of a user-type subject for Azure clusters.

AzureConfigEncryption

Configuration related to config data encryption.

AzureControlPlane

AzureControlPlane represents the control plane configurations.

AzureDatabaseEncryption

Configuration related to application-layer secrets encryption.

AzureDiskTemplate

Configuration for Azure Disks.

AzureK8sVersionInfo

Information about a supported Kubernetes version.

AzureNodeConfig

Parameters that describe the configuration of all node machines on a given node pool.

AzureNodePool

An Anthos node pool running on Azure.

AzureNodePoolAutoscaling

Configuration related to Kubernetes cluster autoscaler.

AzureProxyConfig

Details of a proxy config stored in Azure Key Vault.

AzureServerConfig

AzureServerConfig contains information about a Google Cloud location, such as supported Azure regions and Kubernetes versions.

AzureSshConfig

SSH configuration for Azure resources.

CreateAwsClusterRequest

Request message for AwsClusters.CreateAwsCluster method.

CreateAwsNodePoolRequest

Response message for AwsClusters.CreateAwsNodePool method.

CreateAzureClientRequest

Request message for AzureClusters.CreateAzureClient method.

CreateAzureClusterRequest

Request message for AzureClusters.CreateAzureCluster method.

CreateAzureNodePoolRequest

Response message for AzureClusters.CreateAzureNodePool method.

DeleteAwsClusterRequest

Request message for AwsClusters.DeleteAwsCluster method.

DeleteAwsNodePoolRequest

Request message for AwsClusters.DeleteNodePool method.

DeleteAzureClientRequest

Request message for AzureClusters.DeleteAzureClient method.

DeleteAzureClusterRequest

Request message for Clusters.DeleteAzureCluster method.

DeleteAzureNodePoolRequest

Delete message for AzureClusters.DeleteNodePool method.

Fleet

Fleet related configuration.

GenerateAwsAccessTokenRequest

Request message for AwsClusters.GenerateAwsAccessToken method.

GenerateAwsAccessTokenResponse

Response message for AwsClusters.GenerateAwsAccessToken method.

GenerateAzureAccessTokenRequest

Request message for AzureClusters.GenerateAzureAccessToken method.

GenerateAzureAccessTokenResponse

Response message for AzureClusters.GenerateAzureAccessToken method.

GetAwsClusterRequest

Request message for AwsClusters.GetAwsCluster method.

GetAwsNodePoolRequest

Request message for AwsClusters.GetAwsNodePool method.

GetAwsServerConfigRequest

GetAwsServerConfigRequest gets the server config of GKE cluster on AWS.

GetAzureClientRequest

Request message for AzureClusters.GetAzureClient method.

GetAzureClusterRequest

Request message for AzureClusters.GetAzureCluster method.

GetAzureNodePoolRequest

Request message for AzureClusters.GetAzureNodePool method.

GetAzureServerConfigRequest

GetAzureServerConfigRequest gets the server config of GKE cluster on Azure.

ListAwsClustersRequest

Request message for AwsClusters.ListAwsClusters method.

ListAwsClustersResponse

Response message for AwsClusters.ListAwsClusters method.

ListAwsNodePoolsRequest

Request message for AwsClusters.ListAwsNodePools method.

ListAwsNodePoolsResponse

Response message for AwsClusters.ListAwsNodePools method.

ListAzureClientsRequest

Request message for AzureClusters.ListAzureClients method.

ListAzureClientsResponse

Response message for AzureClusters.ListAzureClients method.

ListAzureClustersRequest

Request message for AzureClusters.ListAzureClusters method.

ListAzureClustersResponse

Response message for AzureClusters.ListAzureClusters method.

ListAzureNodePoolsRequest

Request message for AzureClusters.ListAzureNodePools method.

ListAzureNodePoolsResponse

Response message for AzureClusters.ListAzureNodePools method.

LoggingComponentConfig

Parameters that describe the Logging component configuration in a cluster.

LoggingConfig

Parameters that describe the Logging configuration in a cluster.

MaxPodsConstraint

Constraints applied to pods.

NodeTaint

The taint content for the node taint.

OperationMetadata

Metadata about a long-running operation.

ReplicaPlacement

Configuration for the placement of a control plane replica.

UpdateAwsClusterRequest

Request message for AwsClusters.UpdateAwsCluster method.

UpdateAwsNodePoolRequest

Request message for AwsClusters.UpdateAwsNodePool method.

UpdateAzureClusterRequest

Request message for AzureClusters.UpdateAzureCluster method.

UpdateAzureNodePoolRequest

Request message for AzureClusters.UpdateAzureNodePool method.

WorkloadIdentityConfig

Workload Identity settings.