Module google_api_proto::google::cloud::integrations::v1alpha

Modules

event_execution_details

Nested message and enum types in EventExecutionDetails.

event_execution_snapshot

Nested message and enum types in EventExecutionSnapshot.

execution_info

Nested message and enum types in ExecutionInfo.

failure_policy

Nested message and enum types in FailurePolicy.

success_policy

Nested message and enum types in SuccessPolicy.

task_config

Nested message and enum types in TaskConfig.

task_execution_details

Nested message and enum types in TaskExecutionDetails.

value_type

Nested message and enum types in ValueType.

Structs

AttemptStats

Status for the execution attempt.

BooleanParameterArray

This message only contains a field of boolean array.

ConditionResult

Contains the combined condition calculation results.

DoubleParameterArray

This message only contains a field of double number array.

ErrorDetail

An error, warning, or information message associated with an integration.

EventExecutionDetails

Contains the details of the execution info of this event: this includes the tasks execution details plus the event execution statistics.

EventExecutionSnapshot

Contains the snapshot of the event execution for a given checkpoint.

EventParameter

This message is used for processing and persisting (when applicable) key value pair parameters for each event in the event bus.

ExecutionInfo

Log entry to log execution info for the monitored resource integrations.googleapis.com/IntegrationVersion.

FailurePolicy

Policy that defines the task retry logic and failure type. If no FailurePolicy is defined for a task, all its dependent tasks will not be executed (i.e, a retry_strategy of NONE will be applied).

IntParameterArray

This message only contains a field of integer array.

NextTask

The task that is next in line to be executed, if the condition specified evaluated to true.

StringParameterArray

This message only contains a field of string array.

SuccessPolicy

Policy that dictates the behavior for the task after it completes successfully.

TaskConfig

The task configuration details. This is not the implementation of Task. There might be multiple TaskConfigs for the same Task. (– Next available id: 12 –)

TaskExecutionDetails

Contains the details of the execution of this task.

ValueType

The type of the parameter.

Enums

JsonValidationOption

Options for how to validate json schemas.

Product

Enum Product.