Module google_api_proto::google::cloud::iot::v1

Modules

device_credential

Nested message and enum types in DeviceCredential.

device_manager_client

Generated client implementations.

gateway_list_options

Nested message and enum types in GatewayListOptions.

registry_credential

Nested message and enum types in RegistryCredential.

Structs

BindDeviceToGatewayRequest

Request for BindDeviceToGateway.

BindDeviceToGatewayResponse

Response for BindDeviceToGateway.

CreateDeviceRegistryRequest

Request for CreateDeviceRegistry.

CreateDeviceRequest

Request for CreateDevice.

DeleteDeviceRegistryRequest

Request for DeleteDeviceRegistry.

DeleteDeviceRequest

Request for DeleteDevice.

Device

The device resource.

DeviceConfig

The device configuration. Eventually delivered to devices.

DeviceCredential

A server-stored device credential used for authentication.

DeviceRegistry

A container for a group of devices.

DeviceState

The device state, as reported by the device.

EventNotificationConfig

The configuration for forwarding telemetry events.

GatewayConfig

Gateway-related configuration and state.

GatewayListOptions

Options for limiting the list based on gateway type and associations.

GetDeviceRegistryRequest

Request for GetDeviceRegistry.

GetDeviceRequest

Request for GetDevice.

HttpConfig

The configuration of the HTTP bridge for a device registry.

ListDeviceConfigVersionsRequest

Request for ListDeviceConfigVersions.

ListDeviceConfigVersionsResponse

Response for ListDeviceConfigVersions.

ListDeviceRegistriesRequest

Request for ListDeviceRegistries.

ListDeviceRegistriesResponse

Response for ListDeviceRegistries.

ListDeviceStatesRequest

Request for ListDeviceStates.

ListDeviceStatesResponse

Response for ListDeviceStates.

ListDevicesRequest

Request for ListDevices.

ListDevicesResponse

Response for ListDevices.

ModifyCloudToDeviceConfigRequest

Request for ModifyCloudToDeviceConfig.

MqttConfig

The configuration of MQTT for a device registry.

PublicKeyCertificate

A public key certificate format and data.

PublicKeyCredential

A public key format and data.

RegistryCredential

A server-stored registry credential used to validate device credentials.

SendCommandToDeviceRequest

Request for SendCommandToDevice.

SendCommandToDeviceResponse

Response for SendCommandToDevice.

StateNotificationConfig

The configuration for notification of new states received from the device.

UnbindDeviceFromGatewayRequest

Request for UnbindDeviceFromGateway.

UnbindDeviceFromGatewayResponse

Response for UnbindDeviceFromGateway.

UpdateDeviceRegistryRequest

Request for UpdateDeviceRegistry.

UpdateDeviceRequest

Request for UpdateDevice.

X509CertificateDetails

Details of an X.509 certificate. For informational purposes only.

Enums

GatewayAuthMethod

The gateway authorization/authentication method. This setting determines how Cloud IoT Core authorizes/authenticate devices to access the gateway.

GatewayType

Gateway type.

HttpState

Indicates whether DeviceService (HTTP) is enabled or disabled for the registry. See the field description for details.

LogLevel

Beta Feature

MqttState

Indicates whether an MQTT connection is enabled or disabled. See the field description for details.

PublicKeyCertificateFormat

The supported formats for the public key.

PublicKeyFormat

The supported formats for the public key.