Module google_api_proto::google::cloud::managedkafka::v1

Modules

cluster

Nested message and enum types in Cluster.

managed_kafka_client

Generated client implementations.

rebalance_config

Nested message and enum types in RebalanceConfig.

Structs

AccessConfig

The configuration of access to the Kafka cluster.

CapacityConfig

A capacity configuration of a Kafka cluster.

Cluster

An Apache Kafka cluster deployed in a location.

ConsumerGroup

A Kafka consumer group in a given cluster.

ConsumerPartitionMetadata

Metadata for a consumer group corresponding to a specific partition.

ConsumerTopicMetadata

Metadata for a consumer group corresponding to a specific topic.

CreateClusterRequest

Request for CreateCluster.

CreateTopicRequest

Request for CreateTopic.

DeleteClusterRequest

Request for DeleteCluster.

DeleteConsumerGroupRequest

Request for DeleteConsumerGroup.

DeleteTopicRequest

Request for DeleteTopic.

GcpConfig

Configuration properties for a Kafka cluster deployed to Google Cloud Platform.

GetClusterRequest

Request for GetCluster.

GetConsumerGroupRequest

Request for GetConsumerGroup.

GetTopicRequest

Request for GetTopic.

ListClustersRequest

Request for ListClusters.

ListClustersResponse

Response for ListClusters.

ListConsumerGroupsRequest

Request for ListConsumerGroups.

ListConsumerGroupsResponse

Response for ListConsumerGroups.

ListTopicsRequest

Request for ListTopics.

ListTopicsResponse

Response for ListTopics.

NetworkConfig

The configuration of a Virtual Private Cloud (VPC) network that can access the Kafka cluster.

OperationMetadata

Represents the metadata of the long-running operation.

RebalanceConfig

Defines rebalancing behavior of a Kafka cluster.

Topic

A Kafka topic in a given cluster.

UpdateClusterRequest

Request for UpdateCluster.

UpdateConsumerGroupRequest

Request for UpdateConsumerGroup.

UpdateTopicRequest

Request for UpdateTopic.