Module google_api_proto::google::cloud::migrationcenter::v1::report_summary

Nested message and enum types in ReportSummary.

Modules

chart_data

Nested message and enum types in ChartData.

histogram_chart_data

Nested message and enum types in HistogramChartData.

Structs

AssetAggregateStats

Aggregate statistics for a collection of assets.

ChartData

Describes a collection of data points rendered as a Chart.

ComputeEngineFinding

A set of findings that applies to assets destined for Compute Engine.

GroupFinding

Summary Findings for a specific Group.

GroupPreferenceSetFinding

Summary Findings for a specific Group/PreferenceSet combination.

HistogramChartData

A Histogram Chart shows a distribution of values into buckets, showing a count of values which fall into a bucket.

MachineSeriesAllocation

Represents a data point tracking the count of assets allocated for a specific Machine Series.

SoleTenantFinding

A set of findings that applies to assets destined for Sole-Tenant nodes.

SoleTenantNodeAllocation

Represents the assets allocated to a specific Sole-Tenant node type.

UtilizationChartData

Utilization Chart is a specific type of visualization which displays a metric classified into “Used” and “Free” buckets.

VmwareEngineFinding

A set of findings that applies to assets destined for VMWare Engine.

VmwareNode

A VMWare Engine Node

VmwareNodeAllocation

Represents assets allocated to a specific VMWare Node type.