Module google_api_proto::google::cloud::netapp::v1

Modules

active_directory

Nested message and enum types in ActiveDirectory.

backup

Nested message and enum types in Backup.

backup_policy

Nested message and enum types in BackupPolicy.

backup_vault

Nested message and enum types in BackupVault.

kms_config

Nested message and enum types in KmsConfig.

net_app_client

Generated client implementations.

replication

Nested message and enum types in Replication.

restore_parameters

Nested message and enum types in RestoreParameters.

snapshot

Nested message and enum types in Snapshot.

storage_pool

Nested message and enum types in StoragePool.

volume

Nested message and enum types in Volume.

Structs

ActiveDirectory

ActiveDirectory is the public representation of the active directory config.

Backup

A NetApp Backup.

BackupConfig

BackupConfig contains backup related config on a volume.

BackupPolicy

Backup Policy.

BackupVault

A NetApp BackupVault.

CreateActiveDirectoryRequest

CreateActiveDirectoryRequest for creating an active directory.

CreateBackupPolicyRequest

CreateBackupPolicyRequest creates a backupPolicy.

CreateBackupRequest

CreateBackupRequest creates a backup.

CreateBackupVaultRequest

CreateBackupVaultRequest creates a backup vault.

CreateKmsConfigRequest

CreateKmsConfigRequest creates a KMS Config.

CreateReplicationRequest

CreateReplicationRequest creates a replication.

CreateSnapshotRequest

CreateSnapshotRequest creates a snapshot.

CreateStoragePoolRequest

CreateStoragePoolRequest creates a Storage Pool.

CreateVolumeRequest

Message for creating a Volume

DailySchedule

Make a snapshot every day e.g. at 04:00, 05:20, 23:50

DeleteActiveDirectoryRequest

DeleteActiveDirectoryRequest for deleting a single active directory.

DeleteBackupPolicyRequest

DeleteBackupPolicyRequest deletes a backup policy.

DeleteBackupRequest

DeleteBackupRequest deletes a backup.

DeleteBackupVaultRequest

DeleteBackupVaultRequest deletes a backupVault.

DeleteKmsConfigRequest

DeleteKmsConfigRequest deletes a KMS Config.

DeleteReplicationRequest

DeleteReplicationRequest deletes a replication.

DeleteSnapshotRequest

DeleteSnapshotRequest deletes a snapshot.

DeleteStoragePoolRequest

DeleteStoragePoolRequest deletes a Storage Pool.

DeleteVolumeRequest

Message for deleting a Volume

DestinationVolumeParameters

DestinationVolumeParameters specify input parameters used for creating destination volume.

EncryptVolumesRequest

EncryptVolumesRequest specifies the KMS config to encrypt existing volumes.

ExportPolicy

Defines the export policy for the volume.

GetActiveDirectoryRequest

GetActiveDirectory for getting a single active directory.

GetBackupPolicyRequest

GetBackupPolicyRequest gets the state of a backupPolicy.

GetBackupRequest

GetBackupRequest gets the state of a backup.

GetBackupVaultRequest

GetBackupVaultRequest gets the state of a backupVault.

GetKmsConfigRequest

GetKmsConfigRequest gets a KMS Config.

GetReplicationRequest

GetReplicationRequest gets the state of a replication.

GetSnapshotRequest

GetSnapshotRequest gets the state of a snapshot.

GetStoragePoolRequest

GetStoragePoolRequest gets a Storage Pool.

GetVolumeRequest

Message for getting a Volume

HourlySchedule

Make a snapshot every hour e.g. at 04:00, 05:00, 06:00.

KmsConfig

KmsConfig is the customer managed encryption key(CMEK) configuration.

ListActiveDirectoriesRequest

ListActiveDirectoriesRequest for requesting multiple active directories.

ListActiveDirectoriesResponse

ListActiveDirectoriesResponse contains all the active directories requested.

ListBackupPoliciesRequest

ListBackupPoliciesRequest for requesting multiple backup policies.

ListBackupPoliciesResponse

ListBackupPoliciesResponse contains all the backup policies requested.

ListBackupVaultsRequest

ListBackupVaultsRequest lists backupVaults.

ListBackupVaultsResponse

ListBackupVaultsResponse is the result of ListBackupVaultsRequest.

ListBackupsRequest

ListBackupsRequest lists backups.

ListBackupsResponse

ListBackupsResponse is the result of ListBackupsRequest.

ListKmsConfigsRequest

ListKmsConfigsRequest lists KMS Configs.

ListKmsConfigsResponse

ListKmsConfigsResponse is the response to a ListKmsConfigsRequest.

ListReplicationsRequest

ListReplications lists replications.

ListReplicationsResponse

ListReplicationsResponse is the result of ListReplicationsRequest.

ListSnapshotsRequest

ListSnapshotsRequest lists snapshots.

ListSnapshotsResponse

ListSnapshotsResponse is the result of ListSnapshotsRequest.

ListStoragePoolsRequest

ListStoragePoolsRequest lists Storage Pools.

ListStoragePoolsResponse

ListStoragePoolsResponse is the response to a ListStoragePoolsRequest.

ListVolumesRequest

Message for requesting list of Volumes

ListVolumesResponse

Message for response to listing Volumes

MonthlySchedule

Make a snapshot once a month e.g. at 2nd 04:00, 7th 05:20, 24th 23:50

MountOption

View only mount options for a volume.

OperationMetadata

Represents the metadata of the long-running operation.

Replication

Replication is a nested resource under Volume, that describes a cross-region replication relationship between 2 volumes in different regions.

RestoreParameters

The RestoreParameters if volume is created from a snapshot or backup.

ResumeReplicationRequest

ResumeReplicationRequest resumes a stopped replication.

ReverseReplicationDirectionRequest

ReverseReplicationDirectionRequest reverses direction of replication. Source becomes destination and destination becomes source.

RevertVolumeRequest

RevertVolumeRequest reverts the given volume to the specified snapshot.

SimpleExportPolicyRule

An export policy rule describing various export options.

Snapshot

Snapshot is a point-in-time version of a Volume’s content.

SnapshotPolicy

Snapshot Policy for a volume.

StopReplicationRequest

StopReplicationRequest stops a replication until resumed.

StoragePool

StoragePool is a container for volumes with a service level and capacity. Volumes can be created in a pool of sufficient available capacity. StoragePool capacity is what you are billed for.

TransferStats

TransferStats reports all statistics related to replication transfer.

UpdateActiveDirectoryRequest

UpdateActiveDirectoryRequest for updating an active directory.

UpdateBackupPolicyRequest

UpdateBackupPolicyRequest for updating a backup policy.

UpdateBackupRequest

UpdateBackupRequest updates description and/or labels for a backup.

UpdateBackupVaultRequest

UpdateBackupVaultRequest updates description and/or labels for a backupVault.

UpdateKmsConfigRequest

UpdateKmsConfigRequest updates a KMS Config.

UpdateReplicationRequest

UpdateReplicationRequest updates description and/or labels for a replication.

UpdateSnapshotRequest

UpdateSnapshotRequest updates description and/or labels for a snapshot.

UpdateStoragePoolRequest

UpdateStoragePoolRequest updates a Storage Pool.

UpdateVolumeRequest

Message for updating a Volume

VerifyKmsConfigRequest

VerifyKmsConfigRequest specifies the KMS config to be validated.

VerifyKmsConfigResponse

VerifyKmsConfigResponse contains the information if the config is correctly and error message.

Volume

Volume provides a filesystem that you can mount.

WeeklySchedule

Make a snapshot every week e.g. at Monday 04:00, Wednesday 05:20, Sunday 23:50

Enums

AccessType

AccessType is an enum of all the supported access types for a volume.

EncryptionType

Defined the current volume encryption key source.

Protocols

Protocols is an enum of all the supported network protocols for a volume.

RestrictedAction

Actions to be restricted for a volume.

SecurityStyle

The security style of the volume, can be either UNIX or NTFS.

ServiceLevel

The service levels - Storage Pool, Volumes

SmbSettings

SMBSettings Modifies the behaviour of a SMB volume.