Enum google_api_proto::google::cloud::networkmanagement::v1::drop_info::Cause

#[repr(i32)]
pub enum Cause {

    Unspecified = 0,
    UnknownExternalAddress = 1,
    ForeignIpDisallowed = 2,
    FirewallRule = 3,
    NoRoute = 4,
    RouteBlackhole = 5,
    RouteWrongNetwork = 6,
    PrivateTrafficToInternet = 7,
    PrivateGoogleAccessDisallowed = 8,
    NoExternalAddress = 9,
    UnknownInternalAddress = 10,
    ForwardingRuleMismatch = 11,
    ForwardingRuleRegionMismatch = 25,
    ForwardingRuleNoInstances = 12,
    FirewallBlockingLoadBalancerBackendHealthCheck = 13,
    InstanceNotRunning = 14,
    GkeClusterNotRunning = 27,
    CloudSqlInstanceNotRunning = 28,
    TrafficTypeBlocked = 15,
    GkeMasterUnauthorizedAccess = 16,
    CloudSqlInstanceUnauthorizedAccess = 17,
    DroppedInsideGkeService = 18,
    DroppedInsideCloudSqlService = 19,
    GoogleManagedServiceNoPeering = 20,
    GoogleManagedServiceNoPscEndpoint = 38,
    GkePscEndpointMissing = 36,
    CloudSqlInstanceNoIpAddress = 21,
    GkeControlPlaneRegionMismatch = 30,
    PublicGkeControlPlaneToPrivateDestination = 31,
    GkeControlPlaneNoRoute = 32,
    CloudSqlInstanceNotConfiguredForExternalTraffic = 33,
    PublicCloudSqlInstanceToPrivateDestination = 34,
    CloudSqlInstanceNoRoute = 35,
    CloudFunctionNotActive = 22,
    VpcConnectorNotSet = 23,
    VpcConnectorNotRunning = 24,
    PscConnectionNotAccepted = 26,
    CloudRunRevisionNotReady = 29,
    DroppedInsidePscServiceProducer = 37,
    LoadBalancerHasNoProxySubnet = 39,

}

Drop cause types:

Variants

Unspecified = 0

Cause is unspecified.

UnknownExternalAddress = 1

Destination external address cannot be resolved to a known target. If the address is used in a Google Cloud project, provide the project ID as test input.

ForeignIpDisallowed = 2

A Compute Engine instance can only send or receive a packet with a foreign IP address if ip_forward is enabled.

FirewallRule = 3

Dropped due to a firewall rule, unless allowed due to connection tracking.

NoRoute = 4

Dropped due to no routes.

RouteBlackhole = 5

Dropped due to invalid route. Routeโs next hop is a blackhole.

RouteWrongNetwork = 6

Packet is sent to a wrong (unintended) network. Example: you trace a packet from VM1:Network1 to VM2:Network2, however, the route configured in Network1 sends the packet destined for VM2โs IP addresss to Network3.

PrivateTrafficToInternet = 7

Packet with internal destination address sent to the internet gateway.

PrivateGoogleAccessDisallowed = 8

Instance with only an internal IP address tries to access Google API and services, but private Google access is not enabled.

NoExternalAddress = 9

Instance with only an internal IP address tries to access external hosts, but Cloud NAT is not enabled in the subnet, unless special configurations on a VM allow this connection.

UnknownInternalAddress = 10

Destination internal address cannot be resolved to a known target. If this is a shared VPC scenario, verify if the service project ID is provided as test input. Otherwise, verify if the IP address is being used in the project.

ForwardingRuleMismatch = 11

Forwarding ruleโs protocol and ports do not match the packet header.

ForwardingRuleRegionMismatch = 25

Packet could be dropped because it was sent from a different region to a regional forwarding without global access.

ForwardingRuleNoInstances = 12

Forwarding rule does not have backends configured.

FirewallBlockingLoadBalancerBackendHealthCheck = 13

Firewalls block the health check probes to the backends and cause the backends to be unavailable for traffic from the load balancer. For more details, see Health check firewall rules.

InstanceNotRunning = 14

Packet is sent from or to a Compute Engine instance that is not in a running state.

GkeClusterNotRunning = 27

Packet sent from or to a GKE cluster that is not in running state.

CloudSqlInstanceNotRunning = 28

Packet sent from or to a Cloud SQL instance that is not in running state.

TrafficTypeBlocked = 15

The type of traffic is blocked and the user cannot configure a firewall rule to enable it. See Always blocked traffic for more details.

GkeMasterUnauthorizedAccess = 16

Access to Google Kubernetes Engine cluster masterโs endpoint is not authorized. See Access to the cluster endpoints for more details.

CloudSqlInstanceUnauthorizedAccess = 17

Access to the Cloud SQL instance endpoint is not authorized. See Authorizing with authorized networks for more details.

DroppedInsideGkeService = 18

Packet was dropped inside Google Kubernetes Engine Service.

DroppedInsideCloudSqlService = 19

Packet was dropped inside Cloud SQL Service.

GoogleManagedServiceNoPeering = 20

Packet was dropped because there is no peering between the originating network and the Google Managed Services Network.

GoogleManagedServiceNoPscEndpoint = 38

Packet was dropped because the Google-managed service uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

GkePscEndpointMissing = 36

Packet was dropped because the GKE cluster uses Private Service Connect (PSC), but the PSC endpoint is not found in the project.

CloudSqlInstanceNoIpAddress = 21

Packet was dropped because the Cloud SQL instance has neither a private nor a public IP address.

GkeControlPlaneRegionMismatch = 30

Packet was dropped because a GKE cluster private endpoint is unreachable from a region different from the clusterโs region.

PublicGkeControlPlaneToPrivateDestination = 31

Packet sent from a public GKE cluster control plane to a private IP address.

GkeControlPlaneNoRoute = 32

Packet was dropped because there is no route from a GKE cluster control plane to a destination network.

CloudSqlInstanceNotConfiguredForExternalTraffic = 33

Packet sent from a Cloud SQL instance to an external IP address is not allowed. The Cloud SQL instance is not configured to send packets to external IP addresses.

PublicCloudSqlInstanceToPrivateDestination = 34

Packet sent from a Cloud SQL instance with only a public IP address to a private IP address.

CloudSqlInstanceNoRoute = 35

Packet was dropped because there is no route from a Cloud SQL instance to a destination network.

CloudFunctionNotActive = 22

Packet could be dropped because the Cloud Function is not in an active status.

VpcConnectorNotSet = 23

Packet could be dropped because no VPC connector is set.

VpcConnectorNotRunning = 24

Packet could be dropped because the VPC connector is not in a running state.

PscConnectionNotAccepted = 26

The Private Service Connect endpoint is in a project that is not approved to connect to the service.

CloudRunRevisionNotReady = 29

Packet sent from a Cloud Run revision that is not ready.

DroppedInsidePscServiceProducer = 37

Packet was dropped inside Private Service Connect service producer.

LoadBalancerHasNoProxySubnet = 39

Packet sent to a load balancer, which requires a proxy-only subnet and the subnet is not found.

Implementations

impl Cause

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Cause.

pub fn from_i32(value: i32) -> Option<Cause>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a Cause, or None if value is not a valid variant.

impl Cause

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.