Module google_api_proto::google::cloud::networkmanagement::v1beta1

Modules

abort_info

Nested message and enum types in AbortInfo.

deliver_info

Nested message and enum types in DeliverInfo.

drop_info

Nested message and enum types in DropInfo.

endpoint

Nested message and enum types in Endpoint.

firewall_info

Nested message and enum types in FirewallInfo.

forward_info

Nested message and enum types in ForwardInfo.

google_service_info

Nested message and enum types in GoogleServiceInfo.

load_balancer_backend

Nested message and enum types in LoadBalancerBackend.

load_balancer_info

Nested message and enum types in LoadBalancerInfo.

probing_details

Nested message and enum types in ProbingDetails.

reachability_details

Nested message and enum types in ReachabilityDetails.

reachability_service_client

Generated client implementations.

route_info

Nested message and enum types in RouteInfo.

step

Nested message and enum types in Step.

vpn_tunnel_info

Nested message and enum types in VpnTunnelInfo.

Structs

AbortInfo

Details of the final state “abort” and associated resource.

AppEngineVersionInfo

For display only. Metadata associated with an App Engine version.

CloudFunctionInfo

For display only. Metadata associated with a Cloud Function.

CloudRunRevisionInfo

For display only. Metadata associated with a Cloud Run revision.

CloudSqlInstanceInfo

For display only. Metadata associated with a Cloud SQL instance.

ConnectivityTest

A Connectivity Test for a network reachability analysis.

CreateConnectivityTestRequest

Request for the CreateConnectivityTest method.

DeleteConnectivityTestRequest

Request for the DeleteConnectivityTest method.

DeliverInfo

Details of the final state “deliver” and associated resource.

DropInfo

Details of the final state “drop” and associated resource.

Endpoint

Source or destination of the Connectivity Test.

EndpointInfo

For display only. The specification of the endpoints for the test. EndpointInfo is derived from source and destination Endpoint and validated by the backend data plane model.

FirewallInfo

For display only. Metadata associated with a VPC firewall rule, an implied VPC firewall rule, or a hierarchical firewall policy rule.

ForwardInfo

Details of the final state “forward” and associated resource.

ForwardingRuleInfo

For display only. Metadata associated with a Compute Engine forwarding rule.

GetConnectivityTestRequest

Request for the GetConnectivityTest method.

GkeMasterInfo

For display only. Metadata associated with a Google Kubernetes Engine (GKE) cluster master.

GoogleServiceInfo

For display only. Details of a Google Service sending packets to a VPC network. Although the source IP might be a publicly routable address, some Google Services use special routes within Google production infrastructure to reach Compute Engine Instances. https://cloud.google.com/vpc/docs/routes#special_return_paths

InstanceInfo

For display only. Metadata associated with a Compute Engine instance.

LatencyDistribution

Describes measured latency distribution.

LatencyPercentile

Latency percentile rank and value.

ListConnectivityTestsRequest

Request for the ListConnectivityTests method.

ListConnectivityTestsResponse

Response for the ListConnectivityTests method.

LoadBalancerBackend

For display only. Metadata associated with a specific load balancer backend.

LoadBalancerInfo

For display only. Metadata associated with a load balancer.

NetworkInfo

For display only. Metadata associated with a Compute Engine network.

OperationMetadata

Metadata describing an [Operation][google.longrunning.Operation]

ProbingDetails

Results of active probing from the last run of the test.

ReachabilityDetails

Results of the configuration analysis from the last run of the test.

RerunConnectivityTestRequest

Request for the RerunConnectivityTest method.

RouteInfo

For display only. Metadata associated with a Compute Engine route.

Step

A simulated forwarding path is composed of multiple steps. Each step has a well-defined state and an associated configuration.

Trace

Trace represents one simulated packet forwarding path.

UpdateConnectivityTestRequest

Request for the UpdateConnectivityTest method.

VpcConnectorInfo

For display only. Metadata associated with a VPC connector.

VpnGatewayInfo

For display only. Metadata associated with a Compute Engine VPN gateway.

VpnTunnelInfo

For display only. Metadata associated with a Compute Engine VPN tunnel.

Enums

LoadBalancerType

Type of a load balancer. For more information, see Summary of Google Cloud load balancers.