Module google_api_proto::google::cloud::networkservices::v1

Modules

dep_service_client

Generated client implementations.

endpoint_matcher

Nested message and enum types in EndpointMatcher.

endpoint_policy

Nested message and enum types in EndpointPolicy.

extension_chain

Nested message and enum types in ExtensionChain.

gateway

Nested message and enum types in Gateway.

grpc_route

Nested message and enum types in GrpcRoute.

http_route

Nested message and enum types in HttpRoute.

network_services_client

Generated client implementations.

tcp_route

Nested message and enum types in TcpRoute.

tls_route

Nested message and enum types in TlsRoute.

Structs

CreateEndpointPolicyRequest

Request used with the CreateEndpointPolicy method.

CreateGatewayRequest

Request used by the CreateGateway method.

CreateGrpcRouteRequest

Request used by the CreateGrpcRoute method.

CreateHttpRouteRequest

Request used by the HttpRoute method.

CreateLbRouteExtensionRequest

Message for creating a LbRouteExtension resource.

CreateLbTrafficExtensionRequest

Message for creating a LbTrafficExtension resource.

CreateMeshRequest

Request used by the CreateMesh method.

CreateServiceBindingRequest

Request used by the ServiceBinding method.

CreateTcpRouteRequest

Request used by the TcpRoute method.

CreateTlsRouteRequest

Request used by the TlsRoute method.

DeleteEndpointPolicyRequest

Request used with the DeleteEndpointPolicy method.

DeleteGatewayRequest

Request used by the DeleteGateway method.

DeleteGrpcRouteRequest

Request used by the DeleteGrpcRoute method.

DeleteHttpRouteRequest

Request used by the DeleteHttpRoute method.

DeleteLbRouteExtensionRequest

Message for deleting a LbRouteExtension resource.

DeleteLbTrafficExtensionRequest

Message for deleting a LbTrafficExtension resource.

DeleteMeshRequest

Request used by the DeleteMesh method.

DeleteServiceBindingRequest

Request used by the DeleteServiceBinding method.

DeleteTcpRouteRequest

Request used by the DeleteTcpRoute method.

DeleteTlsRouteRequest

Request used by the DeleteTlsRoute method.

EndpointMatcher

A definition of a matcher that selects endpoints to which the policies should be applied.

EndpointPolicy

EndpointPolicy is a resource that helps apply desired configuration on the endpoints that match specific criteria. For example, this resource can be used to apply “authentication config” an all endpoints that serve on port 8080.

ExtensionChain

A single extension chain wrapper that contains the match conditions and extensions to execute.

Gateway

Gateway represents the configuration for a proxy, typically a load balancer. It captures the ip:port over which the services are exposed by the proxy, along with any policy configurations. Routes have reference to to Gateways to dictate how requests should be routed by this Gateway.

GetEndpointPolicyRequest

Request used with the GetEndpointPolicy method.

GetGatewayRequest

Request used by the GetGateway method.

GetGrpcRouteRequest

Request used by the GetGrpcRoute method.

GetHttpRouteRequest

Request used by the GetHttpRoute method.

GetLbRouteExtensionRequest

Message for getting a LbRouteExtension resource.

GetLbTrafficExtensionRequest

Message for getting a LbTrafficExtension resource.

GetMeshRequest

Request used by the GetMesh method.

GetServiceBindingRequest

Request used by the GetServiceBinding method.

GetTcpRouteRequest

Request used by the GetTcpRoute method.

GetTlsRouteRequest

Request used by the GetTlsRoute method.

GrpcRoute

GrpcRoute is the resource defining how gRPC traffic routed by a Mesh or Gateway resource is routed.

HttpRoute

HttpRoute is the resource defining how HTTP traffic should be routed by a Mesh or Gateway resource.

LbRouteExtension

LbRouteExtension is a resource that lets you control where traffic is routed to for a given request.

LbTrafficExtension

LbTrafficExtension is a resource that lets the extension service modify the headers and payloads of both requests and responses without impacting the choice of backend services or any other security policies associated with the backend service.

ListEndpointPoliciesRequest

Request used with the ListEndpointPolicies method.

ListEndpointPoliciesResponse

Response returned by the ListEndpointPolicies method.

ListGatewaysRequest

Request used with the ListGateways method.

ListGatewaysResponse

Response returned by the ListGateways method.

ListGrpcRoutesRequest

Request used with the ListGrpcRoutes method.

ListGrpcRoutesResponse

Response returned by the ListGrpcRoutes method.

ListHttpRoutesRequest

Request used with the ListHttpRoutes method.

ListHttpRoutesResponse

Response returned by the ListHttpRoutes method.

ListLbRouteExtensionsRequest

Message for requesting list of LbRouteExtension resources.

ListLbRouteExtensionsResponse

Message for response to listing LbRouteExtension resources.

ListLbTrafficExtensionsRequest

Message for requesting list of LbTrafficExtension resources.

ListLbTrafficExtensionsResponse

Message for response to listing LbTrafficExtension resources.

ListMeshesRequest

Request used with the ListMeshes method.

ListMeshesResponse

Response returned by the ListMeshes method.

ListServiceBindingsRequest

Request used with the ListServiceBindings method.

ListServiceBindingsResponse

Response returned by the ListServiceBindings method.

ListTcpRoutesRequest

Request used with the ListTcpRoutes method.

ListTcpRoutesResponse

Response returned by the ListTcpRoutes method.

ListTlsRoutesRequest

Request used with the ListTlsRoutes method.

ListTlsRoutesResponse

Response returned by the ListTlsRoutes method.

Mesh

Mesh represents a logical configuration grouping for workload to workload communication within a service mesh. Routes that point to mesh dictate how requests are routed within this logical mesh boundary.

OperationMetadata

Represents the metadata of the long-running operation.

ServiceBinding

ServiceBinding is the resource that defines a Service Directory Service to be used in a BackendService resource.

TcpRoute

TcpRoute is the resource defining how TCP traffic should be routed by a Mesh/Gateway resource.

TlsRoute

TlsRoute defines how traffic should be routed based on SNI and other matching L3 attributes.

TrafficPortSelector

Specification of a port-based selector.

UpdateEndpointPolicyRequest

Request used with the UpdateEndpointPolicy method.

UpdateGatewayRequest

Request used by the UpdateGateway method.

UpdateGrpcRouteRequest

Request used by the UpdateGrpcRoute method.

UpdateHttpRouteRequest

Request used by the UpdateHttpRoute method.

UpdateLbRouteExtensionRequest

Message for updating a LbRouteExtension resource.

UpdateLbTrafficExtensionRequest

Message for updating a LbTrafficExtension resource.

UpdateMeshRequest

Request used by the UpdateMesh method.

UpdateTcpRouteRequest

Request used by the UpdateTcpRoute method.

UpdateTlsRouteRequest

Request used by the UpdateTlsRoute method.

Enums

EventType

The part of the request or response for which the extension is called.

LoadBalancingScheme

Load balancing schemes supported by the LbTrafficExtension resource and LbRouteExtension resource. For more information, refer to Choosing a load balancer.