Struct google_api_proto::google::cloud::networkservices::v1::network_services_client::NetworkServicesClient

pub struct NetworkServicesClient<T> { /* private fields */ }

Service describing handlers for resources.

Implementations

impl<T> NetworkServicesClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> NetworkServicesClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_endpoint_policies( &mut self, request: impl IntoRequest<ListEndpointPoliciesRequest> ) -> Result<Response<ListEndpointPoliciesResponse>, Status>

Lists EndpointPolicies in a given project and location.

pub async fn get_endpoint_policy( &mut self, request: impl IntoRequest<GetEndpointPolicyRequest> ) -> Result<Response<EndpointPolicy>, Status>

Gets details of a single EndpointPolicy.

pub async fn create_endpoint_policy( &mut self, request: impl IntoRequest<CreateEndpointPolicyRequest> ) -> Result<Response<Operation>, Status>

Creates a new EndpointPolicy in a given project and location.

pub async fn update_endpoint_policy( &mut self, request: impl IntoRequest<UpdateEndpointPolicyRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single EndpointPolicy.

pub async fn delete_endpoint_policy( &mut self, request: impl IntoRequest<DeleteEndpointPolicyRequest> ) -> Result<Response<Operation>, Status>

Deletes a single EndpointPolicy.

pub async fn list_gateways( &mut self, request: impl IntoRequest<ListGatewaysRequest> ) -> Result<Response<ListGatewaysResponse>, Status>

Lists Gateways in a given project and location.

pub async fn get_gateway( &mut self, request: impl IntoRequest<GetGatewayRequest> ) -> Result<Response<Gateway>, Status>

Gets details of a single Gateway.

pub async fn create_gateway( &mut self, request: impl IntoRequest<CreateGatewayRequest> ) -> Result<Response<Operation>, Status>

Creates a new Gateway in a given project and location.

pub async fn update_gateway( &mut self, request: impl IntoRequest<UpdateGatewayRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Gateway.

pub async fn delete_gateway( &mut self, request: impl IntoRequest<DeleteGatewayRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Gateway.

pub async fn list_grpc_routes( &mut self, request: impl IntoRequest<ListGrpcRoutesRequest> ) -> Result<Response<ListGrpcRoutesResponse>, Status>

Lists GrpcRoutes in a given project and location.

pub async fn get_grpc_route( &mut self, request: impl IntoRequest<GetGrpcRouteRequest> ) -> Result<Response<GrpcRoute>, Status>

Gets details of a single GrpcRoute.

pub async fn create_grpc_route( &mut self, request: impl IntoRequest<CreateGrpcRouteRequest> ) -> Result<Response<Operation>, Status>

Creates a new GrpcRoute in a given project and location.

pub async fn update_grpc_route( &mut self, request: impl IntoRequest<UpdateGrpcRouteRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single GrpcRoute.

pub async fn delete_grpc_route( &mut self, request: impl IntoRequest<DeleteGrpcRouteRequest> ) -> Result<Response<Operation>, Status>

Deletes a single GrpcRoute.

pub async fn list_http_routes( &mut self, request: impl IntoRequest<ListHttpRoutesRequest> ) -> Result<Response<ListHttpRoutesResponse>, Status>

Lists HttpRoute in a given project and location.

pub async fn get_http_route( &mut self, request: impl IntoRequest<GetHttpRouteRequest> ) -> Result<Response<HttpRoute>, Status>

Gets details of a single HttpRoute.

pub async fn create_http_route( &mut self, request: impl IntoRequest<CreateHttpRouteRequest> ) -> Result<Response<Operation>, Status>

Creates a new HttpRoute in a given project and location.

pub async fn update_http_route( &mut self, request: impl IntoRequest<UpdateHttpRouteRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single HttpRoute.

pub async fn delete_http_route( &mut self, request: impl IntoRequest<DeleteHttpRouteRequest> ) -> Result<Response<Operation>, Status>

Deletes a single HttpRoute.

pub async fn list_tcp_routes( &mut self, request: impl IntoRequest<ListTcpRoutesRequest> ) -> Result<Response<ListTcpRoutesResponse>, Status>

Lists TcpRoute in a given project and location.

pub async fn get_tcp_route( &mut self, request: impl IntoRequest<GetTcpRouteRequest> ) -> Result<Response<TcpRoute>, Status>

Gets details of a single TcpRoute.

pub async fn create_tcp_route( &mut self, request: impl IntoRequest<CreateTcpRouteRequest> ) -> Result<Response<Operation>, Status>

Creates a new TcpRoute in a given project and location.

pub async fn update_tcp_route( &mut self, request: impl IntoRequest<UpdateTcpRouteRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single TcpRoute.

pub async fn delete_tcp_route( &mut self, request: impl IntoRequest<DeleteTcpRouteRequest> ) -> Result<Response<Operation>, Status>

Deletes a single TcpRoute.

pub async fn list_tls_routes( &mut self, request: impl IntoRequest<ListTlsRoutesRequest> ) -> Result<Response<ListTlsRoutesResponse>, Status>

Lists TlsRoute in a given project and location.

pub async fn get_tls_route( &mut self, request: impl IntoRequest<GetTlsRouteRequest> ) -> Result<Response<TlsRoute>, Status>

Gets details of a single TlsRoute.

pub async fn create_tls_route( &mut self, request: impl IntoRequest<CreateTlsRouteRequest> ) -> Result<Response<Operation>, Status>

Creates a new TlsRoute in a given project and location.

pub async fn update_tls_route( &mut self, request: impl IntoRequest<UpdateTlsRouteRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single TlsRoute.

pub async fn delete_tls_route( &mut self, request: impl IntoRequest<DeleteTlsRouteRequest> ) -> Result<Response<Operation>, Status>

Deletes a single TlsRoute.

pub async fn list_service_bindings( &mut self, request: impl IntoRequest<ListServiceBindingsRequest> ) -> Result<Response<ListServiceBindingsResponse>, Status>

Lists ServiceBinding in a given project and location.

pub async fn get_service_binding( &mut self, request: impl IntoRequest<GetServiceBindingRequest> ) -> Result<Response<ServiceBinding>, Status>

Gets details of a single ServiceBinding.

pub async fn create_service_binding( &mut self, request: impl IntoRequest<CreateServiceBindingRequest> ) -> Result<Response<Operation>, Status>

Creates a new ServiceBinding in a given project and location.

pub async fn delete_service_binding( &mut self, request: impl IntoRequest<DeleteServiceBindingRequest> ) -> Result<Response<Operation>, Status>

Deletes a single ServiceBinding.

pub async fn list_meshes( &mut self, request: impl IntoRequest<ListMeshesRequest> ) -> Result<Response<ListMeshesResponse>, Status>

Lists Meshes in a given project and location.

pub async fn get_mesh( &mut self, request: impl IntoRequest<GetMeshRequest> ) -> Result<Response<Mesh>, Status>

Gets details of a single Mesh.

pub async fn create_mesh( &mut self, request: impl IntoRequest<CreateMeshRequest> ) -> Result<Response<Operation>, Status>

Creates a new Mesh in a given project and location.

pub async fn update_mesh( &mut self, request: impl IntoRequest<UpdateMeshRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single Mesh.

pub async fn delete_mesh( &mut self, request: impl IntoRequest<DeleteMeshRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Mesh.