Module google_api_proto::google::cloud::networkservices::v1beta1

Modules

dep_service_client

Generated client implementations.

endpoint_matcher

Nested message and enum types in EndpointMatcher.

endpoint_policy

Nested message and enum types in EndpointPolicy.

extension_chain

Nested message and enum types in ExtensionChain.

network_services_client

Generated client implementations.

Structs

CreateEndpointPolicyRequest

Request used with the CreateEndpointPolicy method.

CreateLbRouteExtensionRequest

Message for creating a LbRouteExtension resource.

CreateLbTrafficExtensionRequest

Message for creating a LbTrafficExtension resource.

DeleteEndpointPolicyRequest

Request used with the DeleteEndpointPolicy method.

DeleteLbRouteExtensionRequest

Message for deleting a LbRouteExtension resource.

DeleteLbTrafficExtensionRequest

Message for deleting a LbTrafficExtension resource.

EndpointMatcher

A definition of a matcher that selects endpoints to which the policies should be applied.

EndpointPolicy

EndpointPolicy is a resource that helps apply desired configuration on the endpoints that match specific criteria. For example, this resource can be used to apply “authentication config” an all endpoints that serve on port 8080.

ExtensionChain

A single extension chain wrapper that contains the match conditions and extensions to execute.

GetEndpointPolicyRequest

Request used with the GetEndpointPolicy method.

GetLbRouteExtensionRequest

Message for getting a LbRouteExtension resource.

GetLbTrafficExtensionRequest

Message for getting a LbTrafficExtension resource.

LbRouteExtension

LbRouteExtension is a resource that lets you control where traffic is routed to for a given request.

LbTrafficExtension

LbTrafficExtension is a resource that lets the extension service modify the headers and payloads of both requests and responses without impacting the choice of backend services or any other security policies associated with the backend service.

ListEndpointPoliciesRequest

Request used with the ListEndpointPolicies method.

ListEndpointPoliciesResponse

Response returned by the ListEndpointPolicies method.

ListLbRouteExtensionsRequest

Message for requesting list of LbRouteExtension resources.

ListLbRouteExtensionsResponse

Message for response to listing LbRouteExtension resources.

ListLbTrafficExtensionsRequest

Message for requesting list of LbTrafficExtension resources.

ListLbTrafficExtensionsResponse

Message for response to listing LbTrafficExtension resources.

OperationMetadata

Represents the metadata of the long-running operation.

TrafficPortSelector

Specification of a port-based selector.

UpdateEndpointPolicyRequest

Request used with the UpdateEndpointPolicy method.

UpdateLbRouteExtensionRequest

Message for updating a LbRouteExtension resource.

UpdateLbTrafficExtensionRequest

Message for updating a LbTrafficExtension resource.

Enums

LoadBalancingScheme

Load balancing schemes supported by the LbTrafficExtension resource and LbRouteExtension resource. The valid values are: INTERNAL_MANAGED, EXTERNAL_MANAGED.