Struct google_api_proto::google::cloud::notebooks::v1::notebook_service_client::NotebookServiceClient

pub struct NotebookServiceClient<T> { /* private fields */ }

API v1 service for Cloud AI Platform Notebooks.

Implementations

impl<T> NotebookServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> NotebookServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_instances( &mut self, request: impl IntoRequest<ListInstancesRequest> ) -> Result<Response<ListInstancesResponse>, Status>

Lists instances in a given project and location.

pub async fn get_instance( &mut self, request: impl IntoRequest<GetInstanceRequest> ) -> Result<Response<Instance>, Status>

Gets details of a single Instance.

pub async fn create_instance( &mut self, request: impl IntoRequest<CreateInstanceRequest> ) -> Result<Response<Operation>, Status>

Creates a new Instance in a given project and location.

pub async fn register_instance( &mut self, request: impl IntoRequest<RegisterInstanceRequest> ) -> Result<Response<Operation>, Status>

Registers an existing legacy notebook instance to the Notebooks API server. Legacy instances are instances created with the legacy Compute Engine calls. They are not manageable by the Notebooks API out of the box. This call makes these instances manageable by the Notebooks API.

pub async fn set_instance_accelerator( &mut self, request: impl IntoRequest<SetInstanceAcceleratorRequest> ) -> Result<Response<Operation>, Status>

Updates the guest accelerators of a single Instance.

pub async fn set_instance_machine_type( &mut self, request: impl IntoRequest<SetInstanceMachineTypeRequest> ) -> Result<Response<Operation>, Status>

Updates the machine type of a single Instance.

pub async fn update_instance_config( &mut self, request: impl IntoRequest<UpdateInstanceConfigRequest> ) -> Result<Response<Operation>, Status>

Update Notebook Instance configurations.

pub async fn update_shielded_instance_config( &mut self, request: impl IntoRequest<UpdateShieldedInstanceConfigRequest> ) -> Result<Response<Operation>, Status>

Updates the Shielded instance configuration of a single Instance.

pub async fn set_instance_labels( &mut self, request: impl IntoRequest<SetInstanceLabelsRequest> ) -> Result<Response<Operation>, Status>

Replaces all the labels of an Instance.

pub async fn update_instance_metadata_items( &mut self, request: impl IntoRequest<UpdateInstanceMetadataItemsRequest> ) -> Result<Response<UpdateInstanceMetadataItemsResponse>, Status>

Add/update metadata items for an instance.

pub async fn delete_instance( &mut self, request: impl IntoRequest<DeleteInstanceRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Instance.

pub async fn start_instance( &mut self, request: impl IntoRequest<StartInstanceRequest> ) -> Result<Response<Operation>, Status>

Starts a notebook instance.

pub async fn stop_instance( &mut self, request: impl IntoRequest<StopInstanceRequest> ) -> Result<Response<Operation>, Status>

Stops a notebook instance.

pub async fn reset_instance( &mut self, request: impl IntoRequest<ResetInstanceRequest> ) -> Result<Response<Operation>, Status>

Resets a notebook instance.

pub async fn report_instance_info( &mut self, request: impl IntoRequest<ReportInstanceInfoRequest> ) -> Result<Response<Operation>, Status>

Allows notebook instances to report their latest instance information to the Notebooks API server. The server will merge the reported information to the instance metadata store. Do not use this method directly.

pub async fn is_instance_upgradeable( &mut self, request: impl IntoRequest<IsInstanceUpgradeableRequest> ) -> Result<Response<IsInstanceUpgradeableResponse>, Status>

Check if a notebook instance is upgradable.

pub async fn get_instance_health( &mut self, request: impl IntoRequest<GetInstanceHealthRequest> ) -> Result<Response<GetInstanceHealthResponse>, Status>

Check if a notebook instance is healthy.

pub async fn upgrade_instance( &mut self, request: impl IntoRequest<UpgradeInstanceRequest> ) -> Result<Response<Operation>, Status>

Upgrades a notebook instance to the latest version.

pub async fn rollback_instance( &mut self, request: impl IntoRequest<RollbackInstanceRequest> ) -> Result<Response<Operation>, Status>

Rollbacks a notebook instance to the previous version.

pub async fn diagnose_instance( &mut self, request: impl IntoRequest<DiagnoseInstanceRequest> ) -> Result<Response<Operation>, Status>

Creates a Diagnostic File and runs Diagnostic Tool given an Instance.

pub async fn upgrade_instance_internal( &mut self, request: impl IntoRequest<UpgradeInstanceInternalRequest> ) -> Result<Response<Operation>, Status>

Allows notebook instances to call this endpoint to upgrade themselves. Do not use this method directly.

pub async fn list_environments( &mut self, request: impl IntoRequest<ListEnvironmentsRequest> ) -> Result<Response<ListEnvironmentsResponse>, Status>

Lists environments in a project.

pub async fn get_environment( &mut self, request: impl IntoRequest<GetEnvironmentRequest> ) -> Result<Response<Environment>, Status>

Gets details of a single Environment.

pub async fn create_environment( &mut self, request: impl IntoRequest<CreateEnvironmentRequest> ) -> Result<Response<Operation>, Status>

Creates a new Environment.

pub async fn delete_environment( &mut self, request: impl IntoRequest<DeleteEnvironmentRequest> ) -> Result<Response<Operation>, Status>

Deletes a single Environment.

pub async fn list_schedules( &mut self, request: impl IntoRequest<ListSchedulesRequest> ) -> Result<Response<ListSchedulesResponse>, Status>

Lists schedules in a given project and location.

pub async fn get_schedule( &mut self, request: impl IntoRequest<GetScheduleRequest> ) -> Result<Response<Schedule>, Status>

Gets details of schedule

pub async fn delete_schedule( &mut self, request: impl IntoRequest<DeleteScheduleRequest> ) -> Result<Response<Operation>, Status>

Deletes schedule and all underlying jobs

pub async fn create_schedule( &mut self, request: impl IntoRequest<CreateScheduleRequest> ) -> Result<Response<Operation>, Status>

Creates a new Scheduled Notebook in a given project and location.

pub async fn trigger_schedule( &mut self, request: impl IntoRequest<TriggerScheduleRequest> ) -> Result<Response<Operation>, Status>

Triggers execution of an existing schedule.

pub async fn list_executions( &mut self, request: impl IntoRequest<ListExecutionsRequest> ) -> Result<Response<ListExecutionsResponse>, Status>

Lists executions in a given project and location

pub async fn get_execution( &mut self, request: impl IntoRequest<GetExecutionRequest> ) -> Result<Response<Execution>, Status>

Gets details of executions

pub async fn delete_execution( &mut self, request: impl IntoRequest<DeleteExecutionRequest> ) -> Result<Response<Operation>, Status>

Deletes execution

pub async fn create_execution( &mut self, request: impl IntoRequest<CreateExecutionRequest> ) -> Result<Response<Operation>, Status>

Creates a new Execution in a given project and location.