Module google_api_proto::google::cloud::orchestration::airflow::service::v1beta1

Modules

check_upgrade_response

Nested message and enum types in CheckUpgradeResponse.

environment

Nested message and enum types in Environment.

environment_config

Nested message and enum types in EnvironmentConfig.

environments_client

Generated client implementations.

image_versions_client

Generated client implementations.

master_authorized_networks_config

Nested message and enum types in MasterAuthorizedNetworksConfig.

networking_config

Nested message and enum types in NetworkingConfig.

operation_metadata

Nested message and enum types in OperationMetadata.

poll_airflow_command_response

Nested message and enum types in PollAirflowCommandResponse.

web_server_network_access_control

Nested message and enum types in WebServerNetworkAccessControl.

workloads_config

Nested message and enum types in WorkloadsConfig.

Structs

CheckUpgradeRequest

Request to check whether image upgrade will succeed.

CheckUpgradeResponse

Message containing information about the result of an upgrade check operation.

CloudDataLineageIntegration

Configuration for Cloud Data Lineage integration.

CreateEnvironmentRequest

Create a new environment.

DatabaseConfig

The configuration of Cloud SQL instance that is used by the Apache Airflow software.

DatabaseFailoverRequest

Request to trigger database failover (only for highly resilient environments).

DatabaseFailoverResponse

Response for DatabaseFailoverRequest.

DeleteEnvironmentRequest

Delete an environment.

EncryptionConfig

The encryption options for the Cloud Composer environment and its dependencies. Supported for Cloud Composer environments in versions composer-1..-airflow-..*.

Environment

An environment for running orchestration tasks.

EnvironmentConfig

Configuration information for an environment.

ExecuteAirflowCommandRequest

Execute Airflow Command request.

ExecuteAirflowCommandResponse

Response to ExecuteAirflowCommandRequest.

FetchDatabasePropertiesRequest

Request to fetch properties of environment’s database.

FetchDatabasePropertiesResponse

Response for FetchDatabasePropertiesRequest.

GetEnvironmentRequest

Get an environment.

ImageVersion

Image Version information

IpAllocationPolicy

Configuration for controlling how IPs are allocated in the GKE cluster.

ListEnvironmentsRequest

List environments in a project and location.

ListEnvironmentsResponse

The environments in a project and location.

ListImageVersionsRequest

List ImageVersions in a project and location.

ListImageVersionsResponse

The ImageVersions in a project and location.

LoadSnapshotRequest

Request to load a snapshot into a Cloud Composer environment.

LoadSnapshotResponse

Response to LoadSnapshotRequest.

MaintenanceWindow

The configuration settings for Cloud Composer maintenance window. The following example:

MasterAuthorizedNetworksConfig

Configuration options for the master authorized networks feature. Enabled master authorized networks will disallow all external traffic to access Kubernetes master through HTTPS except traffic from the given CIDR blocks, Google Compute Engine Public IPs and Google Prod IPs.

NetworkingConfig

Configuration options for networking connections in the Composer 2 environment.

NodeConfig

The configuration information for the Kubernetes Engine nodes running the Apache Airflow software.

OperationMetadata

Metadata describing an operation.

PollAirflowCommandRequest

Poll Airflow Command request.

PollAirflowCommandResponse

Response to PollAirflowCommandRequest.

PrivateClusterConfig

Configuration options for the private GKE cluster in a Cloud Composer environment.

PrivateEnvironmentConfig

The configuration information for configuring a Private IP Cloud Composer environment.

RecoveryConfig

The Recovery settings of an environment.

RestartWebServerRequest

Restart Airflow web server.

SaveSnapshotRequest

Request to create a snapshot of a Cloud Composer environment.

SaveSnapshotResponse

Response to SaveSnapshotRequest.

ScheduledSnapshotsConfig

The configuration for scheduled snapshot creation mechanism.

SoftwareConfig

Specifies the selection and configuration of software inside the environment.

StopAirflowCommandRequest

Stop Airflow Command request.

StopAirflowCommandResponse

Response to StopAirflowCommandRequest.

UpdateEnvironmentRequest

Update an environment.

WebServerConfig

The configuration settings for the Airflow web server App Engine instance. Supported for Cloud Composer environments in versions composer-1..-airflow-..*.

WebServerNetworkAccessControl

Network-level access control policy for the Airflow web server.

WorkloadsConfig

The Kubernetes workloads configuration for GKE cluster associated with the Cloud Composer environment. Supported for Cloud Composer environments in versions composer-2..-airflow-..* and newer.