Module google_api_proto::google::cloud::orgpolicy::v2

Modules

constraint

Nested message and enum types in Constraint.

custom_constraint

Nested message and enum types in CustomConstraint.

org_policy_client

Generated client implementations.

policy_spec

Nested message and enum types in PolicySpec.

Structs

AlternatePolicySpec

Similar to PolicySpec but with an extra ‘launch’ field for launch reference. The PolicySpec here is specific for dry-run/darklaunch.

Constraint

A constraint describes a way to restrict resource’s configuration. For example, you could enforce a constraint that controls which Google Cloud services can be activated across an organization, or whether a Compute Engine instance can have serial port connections established. Constraints can be configured by the organization policy administrator to fit the needs of the organization by setting a policy that includes constraints at different locations in the organization’s resource hierarchy. Policies are inherited down the resource hierarchy from higher levels, but can also be overridden. For details about the inheritance rules please read about [policies][google.cloud.OrgPolicy.v2.Policy].

CreateCustomConstraintRequest

The request sent to the [CreateCustomConstraintRequest] [google.cloud.orgpolicy.v2.OrgPolicy.CreateCustomConstraint] method.

CreatePolicyRequest

The request sent to the [CreatePolicyRequest] [google.cloud.orgpolicy.v2.OrgPolicy.CreatePolicy] method.

CustomConstraint

A custom constraint defined by customers which can only be applied to the given resource types and organization.

DeleteCustomConstraintRequest

The request sent to the [DeleteCustomConstraint] [google.cloud.orgpolicy.v2.OrgPolicy.DeleteCustomConstraint] method.

DeletePolicyRequest

The request sent to the [DeletePolicy] [google.cloud.orgpolicy.v2.OrgPolicy.DeletePolicy] method.

GetCustomConstraintRequest

The request sent to the [GetCustomConstraint] [google.cloud.orgpolicy.v2.OrgPolicy.GetCustomConstraint] method.

GetEffectivePolicyRequest

The request sent to the [GetEffectivePolicy] [google.cloud.orgpolicy.v2.OrgPolicy.GetEffectivePolicy] method.

GetPolicyRequest

The request sent to the [GetPolicy] [google.cloud.orgpolicy.v2.OrgPolicy.GetPolicy] method.

ListConstraintsRequest

The request sent to the [ListConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListConstraints] method.

ListConstraintsResponse

The response returned from the [ListConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListConstraints] method.

ListCustomConstraintsRequest

The request sent to the [ListCustomConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListCustomConstraints] method.

ListCustomConstraintsResponse

The response returned from the [ListCustomConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListCustomConstraints] method. It will be empty if no custom constraints are set on the organization resource.

ListPoliciesRequest

The request sent to the [ListPolicies] [google.cloud.orgpolicy.v2.OrgPolicy.ListPolicies] method.

ListPoliciesResponse

The response returned from the [ListPolicies] [google.cloud.orgpolicy.v2.OrgPolicy.ListPolicies] method. It will be empty if no policies are set on the resource.

Policy

Defines an organization policy which is used to specify constraints for configurations of Google Cloud resources.

PolicySpec

Defines a Google Cloud policy specification which is used to specify constraints for configurations of Google Cloud resources.

UpdateCustomConstraintRequest

The request sent to the [UpdateCustomConstraintRequest] [google.cloud.orgpolicy.v2.OrgPolicy.UpdateCustomConstraint] method.

UpdatePolicyRequest

The request sent to the [UpdatePolicyRequest] [google.cloud.orgpolicy.v2.OrgPolicy.UpdatePolicy] method.