Module google_api_proto::google::cloud::oslogin::v1beta

Modules

os_login_service_client

Generated client implementations.

security_key

Nested message and enum types in SecurityKey.

Structs

CreateSshPublicKeyRequest

A request message for creating an SSH public key.

DeletePosixAccountRequest

A request message for deleting a POSIX account entry.

DeleteSshPublicKeyRequest

A request message for deleting an SSH public key.

GetLoginProfileRequest

A request message for retrieving the login profile information for a user.

GetSshPublicKeyRequest

A request message for retrieving an SSH public key.

ImportSshPublicKeyRequest

A request message for importing an SSH public key.

ImportSshPublicKeyResponse

A response message for importing an SSH public key.

LoginProfile

The user profile information used for logging in to a virtual machine on Google Compute Engine.

SecurityKey

The credential information for a Google registered security key.

SignSshPublicKeyRequest

SignSshPublicKeyResponse

UniversalTwoFactor

Security key information specific to the U2F protocol.

UpdateSshPublicKeyRequest

A request message for updating an SSH public key.

WebAuthn

Security key information specific to the Web Authentication protocol.

Enums

LoginProfileView

The login profile view limits the user content retrieved.