Module google_api_proto::google::cloud::paymentgateway::issuerswitch::accountmanager::v1

Modules

account_manager_participant

Nested message and enum types in AccountManagerParticipant.

account_manager_transaction_info

Nested message and enum types in AccountManagerTransactionInfo.

account_manager_transaction_reconciliation_info

Nested message and enum types in AccountManagerTransactionReconciliationInfo.

account_manager_transactions_client

Generated client implementations.

managed_account

Nested message and enum types in ManagedAccount.

managed_accounts_client

Generated client implementations.

Structs

AccountManagerMerchantInfo

A merchant in a transaction processed by the account manager.

AccountManagerParticipant

A participant in a transaction processed by the account manager.

AccountManagerSettlementParticipant

A participant in a payment settlement transaction processed by the account manager. The participant could either be the payer or the payee in the transaction.

AccountManagerTransaction

A transaction processed by the account manager.

AccountManagerTransactionInfo

Information about a transaction processed by the account manager.

AccountManagerTransactionReconciliationInfo

Reconciliation information for a transaction processed by account manager.

BatchReconcileAccountManagerTransactionsRequest

Request for the BatchReconcileAccountManagerTransactions method.

BatchReconcileAccountManagerTransactionsResponse

Response for the BatchReconcileAccountManagerTransactions method.

BatchReconcileManagedAccountBalanceRequest

Request for the BatchReconcileManagedAccountBalance method.

BatchReconcileManagedAccountBalanceResponse

Response for the BatchReconcileManagedAccountBalance method.

ExportAccountManagerTransactionsRequest

Request for the ExportAccountManagerTransactions method.

GetManagedAccountRequest

Request for the GetManagedAccount method.

ListAccountManagerTransactionsRequest

Request for the ListAccountManagerTransactions method. Callers can request for transactions to be filtered by the given filter criteria and specified pagination parameters.

ListAccountManagerTransactionsResponse

Response for the ListAccountManagerTransactions method.

ManagedAccount

Entity representing an account managed by the account manager.

ReconcileAccountManagerTransactionsRequest

Reconciliation request for an account manager transaction.

ReconcileManagedAccountBalanceRequest

Reconciliation request for an account balance.

Enums

AccountManagerTransactionType

The type of a account manager transaction. Every transaction processed by the account manager will be one of these transaction types.