Module google_api_proto::google::cloud::recaptchaenterprise::v1

Modules

account_defender_assessment

Nested message and enum types in AccountDefenderAssessment.

annotate_assessment_request

Nested message and enum types in AnnotateAssessmentRequest.

key

Nested message and enum types in Key.

recaptcha_enterprise_service_client

Generated client implementations.

risk_analysis

Nested message and enum types in RiskAnalysis.

testing_options

Nested message and enum types in TestingOptions.

token_properties

Nested message and enum types in TokenProperties.

web_key_settings

Nested message and enum types in WebKeySettings.

Structs

AccountDefenderAssessment

Account Defender risk assessment.

AndroidKeySettings

Settings specific to keys that can be used by Android apps.

AnnotateAssessmentRequest

The request message to annotate an Assessment.

AnnotateAssessmentResponse

Empty response for AnnotateAssessment.

Assessment

A recaptcha assessment resource.

ChallengeMetrics

Metrics related to challenges.

CreateAssessmentRequest

The create assessment request message.

CreateKeyRequest

The create key request message.

DeleteKeyRequest

The delete key request message.

Event

GetKeyRequest

The get key request message.

GetMetricsRequest

The get metrics request message.

IosKeySettings

Settings specific to keys that can be used by iOS apps.

Key

A key used to identify and configure applications (web and/or mobile) that use reCAPTCHA Enterprise.

ListKeysRequest

The list keys request message.

ListKeysResponse

Response to request to list keys in a project.

ListRelatedAccountGroupMembershipsRequest

The request message to list memberships in a related account group.

ListRelatedAccountGroupMembershipsResponse

The response to a ListRelatedAccountGroupMemberships call.

ListRelatedAccountGroupsRequest

The request message to list related account groups.

ListRelatedAccountGroupsResponse

The response to a ListRelatedAccountGroups call.

Metrics

Metrics for a single Key.

MigrateKeyRequest

The migrate key request message.

RelatedAccountGroup

A group of related accounts.

RelatedAccountGroupMembership

A membership in a group of related accounts.

RiskAnalysis

Risk analysis result for an event.

ScoreDistribution

Score distribution.

ScoreMetrics

Metrics related to scoring.

SearchRelatedAccountGroupMembershipsRequest

The request message to search related account group memberships.

SearchRelatedAccountGroupMembershipsResponse

The response to a SearchRelatedAccountGroupMemberships call.

TestingOptions

Options for user acceptance testing.

TokenProperties

UpdateKeyRequest

The update key request message.

WebKeySettings

Settings specific to keys that can be used by websites.