Module google_api_proto::google::cloud::redis::v1beta1

Modules

cloud_redis_client

Generated client implementations.

failover_instance_request

Nested message and enum types in FailoverInstanceRequest.

input_config

Nested message and enum types in InputConfig.

instance

Nested message and enum types in Instance.

output_config

Nested message and enum types in OutputConfig.

reschedule_maintenance_request

Nested message and enum types in RescheduleMaintenanceRequest.

Structs

CreateInstanceRequest

Request for [CreateInstance][google.cloud.redis.v1beta1.CloudRedis.CreateInstance].

DeleteInstanceRequest

Request for [DeleteInstance][google.cloud.redis.v1beta1.CloudRedis.DeleteInstance].

ExportInstanceRequest

Request for [Export][google.cloud.redis.v1beta1.CloudRedis.ExportInstance].

FailoverInstanceRequest

Request for [Failover][google.cloud.redis.v1beta1.CloudRedis.FailoverInstance].

GcsDestination

The Cloud Storage location for the output content

GcsSource

The Cloud Storage location for the input content

GetInstanceAuthStringRequest

Request for [GetInstanceAuthString][google.cloud.redis.v1beta1.CloudRedis.GetInstanceAuthString].

GetInstanceRequest

Request for [GetInstance][google.cloud.redis.v1beta1.CloudRedis.GetInstance].

ImportInstanceRequest

Request for [Import][google.cloud.redis.v1beta1.CloudRedis.ImportInstance].

InputConfig

The input content

Instance

A Google Cloud Redis instance.

InstanceAuthString

Instance AUTH string details.

ListInstancesRequest

Request for [ListInstances][google.cloud.redis.v1beta1.CloudRedis.ListInstances].

ListInstancesResponse

Response for [ListInstances][google.cloud.redis.v1beta1.CloudRedis.ListInstances].

LocationMetadata

This location metadata represents additional configuration options for a given location where a Redis instance may be created. All fields are output only. It is returned as content of the google.cloud.location.Location.metadata field.

MaintenancePolicy

Maintenance policy for an instance.

MaintenanceSchedule

Upcoming maintenance schedule. If no maintenance is scheduled, fields are not populated.

NodeInfo

Node specific properties.

OutputConfig

The output content

RescheduleMaintenanceRequest

Request for [RescheduleMaintenance][google.cloud.redis.v1beta1.CloudRedis.RescheduleMaintenance].

TlsCertificate

TlsCertificate Resource

UpdateInstanceRequest

Request for [UpdateInstance][google.cloud.redis.v1beta1.CloudRedis.UpdateInstance].

UpgradeInstanceRequest

Request for [UpgradeInstance][google.cloud.redis.v1beta1.CloudRedis.UpgradeInstance].

WeeklyMaintenanceWindow

Time window in which disruptive maintenance updates occur. Non-disruptive updates can occur inside or outside this window.

ZoneMetadata

Defines specific information for a particular zone. Currently empty and reserved for future use only.