Module google_api_proto::google::cloud::retail::v2

Modules

big_query_source

Nested message and enum types in BigQuerySource.

catalog_attribute

Nested message and enum types in CatalogAttribute.

catalog_service_client

Generated client implementations.

collect_user_event_request

Nested message and enum types in CollectUserEventRequest.

complete_query_response

Nested message and enum types in CompleteQueryResponse.

completion_data_input_config

Nested message and enum types in CompletionDataInputConfig.

completion_service_client

Generated client implementations.

condition

Nested message and enum types in Condition.

control

Nested message and enum types in Control.

control_service_client

Generated client implementations.

experiment_info

Nested message and enum types in ExperimentInfo.

import_errors_config

Nested message and enum types in ImportErrorsConfig.

import_products_request

Nested message and enum types in ImportProductsRequest.

interval

Nested message and enum types in Interval.

model

Nested message and enum types in Model.

model_service_client

Generated client implementations.

predict_response

Nested message and enum types in PredictResponse.

prediction_service_client

Generated client implementations.

price_info

Nested message and enum types in PriceInfo.

product

Nested message and enum types in Product.

product_input_config

Nested message and enum types in ProductInputConfig.

product_service_client

Generated client implementations.

rejoin_user_events_request

Nested message and enum types in RejoinUserEventsRequest.

rule

Nested message and enum types in Rule.

search_request

Nested message and enum types in SearchRequest.

search_response

Nested message and enum types in SearchResponse.

search_service_client

Generated client implementations.

serving_config

Nested message and enum types in ServingConfig.

serving_config_service_client

Generated client implementations.

user_event_input_config

Nested message and enum types in UserEventInputConfig.

user_event_service_client

Generated client implementations.

Structs

AddCatalogAttributeRequest

Request for [CatalogService.AddCatalogAttribute][google.cloud.retail.v2.CatalogService.AddCatalogAttribute] method.

AddControlRequest

Request for AddControl method.

AddFulfillmentPlacesMetadata

Metadata related to the progress of the AddFulfillmentPlaces operation. Currently empty because there is no meaningful metadata populated from the [ProductService.AddFulfillmentPlaces][google.cloud.retail.v2.ProductService.AddFulfillmentPlaces] method.

AddFulfillmentPlacesRequest

Request message for [ProductService.AddFulfillmentPlaces][google.cloud.retail.v2.ProductService.AddFulfillmentPlaces] method.

AddFulfillmentPlacesResponse

Response of the AddFulfillmentPlacesRequest. Currently empty because there is no meaningful response populated from the [ProductService.AddFulfillmentPlaces][google.cloud.retail.v2.ProductService.AddFulfillmentPlaces] method.

AddLocalInventoriesMetadata

Metadata related to the progress of the AddLocalInventories operation. Currently empty because there is no meaningful metadata populated from the [ProductService.AddLocalInventories][google.cloud.retail.v2.ProductService.AddLocalInventories] method.

AddLocalInventoriesRequest

Request message for [ProductService.AddLocalInventories][google.cloud.retail.v2.ProductService.AddLocalInventories] method.

AddLocalInventoriesResponse

Response of the [ProductService.AddLocalInventories][google.cloud.retail.v2.ProductService.AddLocalInventories] API. Currently empty because there is no meaningful response populated from the [ProductService.AddLocalInventories][google.cloud.retail.v2.ProductService.AddLocalInventories] method.

AttributesConfig

Catalog level attribute config.

Audience

An intended audience of the [Product][google.cloud.retail.v2.Product] for whom it’s sold.

BigQuerySource

BigQuery source import data from.

Catalog

The catalog configuration.

CatalogAttribute

Catalog level attribute config for an attribute. For example, if customers want to enable/disable facet for a specific attribute.

CollectUserEventRequest

Request message for CollectUserEvent method.

ColorInfo

The color information of a [Product][google.cloud.retail.v2.Product].

CompleteQueryRequest

Autocomplete parameters.

CompleteQueryResponse

Response of the autocomplete query.

CompletionConfig

Catalog level autocomplete config for customers to customize autocomplete feature’s settings.

CompletionDataInputConfig

The input config source for completion data.

CompletionDetail

Detailed completion information including completion attribution token and clicked completion info.

Condition

Metadata that is used to define a condition that triggers an action. A valid condition must specify at least one of ‘query_terms’ or ‘products_filter’. If multiple fields are specified, the condition is met if all the fields are satisfied e.g. if a set of query terms and product_filter are set, then only items matching the product_filter for requests with a query matching the query terms wil get boosted.

Control

Configures dynamic metadata that can be linked to a [ServingConfig][google.cloud.retail.v2.ServingConfig] and affect search or recommendation results at serving time.

CreateControlRequest

Request for CreateControl method.

CreateModelMetadata

Metadata associated with a create operation.

CreateModelRequest

Request for creating a model.

CreateProductRequest

Request message for [ProductService.CreateProduct][google.cloud.retail.v2.ProductService.CreateProduct] method.

CreateServingConfigRequest

Request for CreateServingConfig method.

CustomAttribute

A custom attribute that is not explicitly modeled in [Product][google.cloud.retail.v2.Product].

DeleteControlRequest

Request for DeleteControl method.

DeleteModelRequest

Request for deleting a model.

DeleteProductRequest

Request message for [ProductService.DeleteProduct][google.cloud.retail.v2.ProductService.DeleteProduct] method.

DeleteServingConfigRequest

Request for DeleteServingConfig method.

ExperimentInfo

Metadata for active A/B testing [Experiments][].

FulfillmentInfo

Fulfillment information, such as the store IDs for in-store pickup or region IDs for different shipping methods.

GcsSource

Google Cloud Storage location for input content.

GetAttributesConfigRequest

Request for [CatalogService.GetAttributesConfig][google.cloud.retail.v2.CatalogService.GetAttributesConfig] method.

GetCompletionConfigRequest

Request for [CatalogService.GetCompletionConfig][google.cloud.retail.v2.CatalogService.GetCompletionConfig] method.

GetControlRequest

Request for GetControl method.

GetDefaultBranchRequest

Request message to show which branch is currently the default branch.

GetDefaultBranchResponse

Response message of [CatalogService.GetDefaultBranch][google.cloud.retail.v2.CatalogService.GetDefaultBranch].

GetModelRequest

Request for getting a model.

GetProductRequest

Request message for [ProductService.GetProduct][google.cloud.retail.v2.ProductService.GetProduct] method.

GetServingConfigRequest

Request for GetServingConfig method.

Image

[Product][google.cloud.retail.v2.Product] image. Recommendations AI and Retail Search do not use product images to improve prediction and search results. However, product images can be returned in results, and are shown in prediction or search previews in the console.

ImportCompletionDataRequest

Request message for ImportCompletionData methods.

ImportCompletionDataResponse

Response of the [ImportCompletionDataRequest][google.cloud.retail.v2.ImportCompletionDataRequest]. If the long running operation is done, this message is returned by the google.longrunning.Operations.response field if the operation is successful.

ImportErrorsConfig

Configuration of destination for Import related errors.

ImportMetadata

Metadata related to the progress of the Import operation. This is returned by the google.longrunning.Operation.metadata field.

ImportProductsRequest

Request message for Import methods.

ImportProductsResponse

Response of the [ImportProductsRequest][google.cloud.retail.v2.ImportProductsRequest]. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportUserEventsRequest

Request message for the ImportUserEvents request.

ImportUserEventsResponse

Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

Interval

A floating point interval.

ListCatalogsRequest

Request for [CatalogService.ListCatalogs][google.cloud.retail.v2.CatalogService.ListCatalogs] method.

ListCatalogsResponse

Response for [CatalogService.ListCatalogs][google.cloud.retail.v2.CatalogService.ListCatalogs] method.

ListControlsRequest

Request for ListControls method.

ListControlsResponse

Response for ListControls method.

ListModelsRequest

Request for listing models associated with a resource.

ListModelsResponse

Response to a ListModelRequest.

ListProductsRequest

Request message for [ProductService.ListProducts][google.cloud.retail.v2.ProductService.ListProducts] method.

ListProductsResponse

Response message for [ProductService.ListProducts][google.cloud.retail.v2.ProductService.ListProducts] method.

ListServingConfigsRequest

Request for ListServingConfigs method.

ListServingConfigsResponse

Response for ListServingConfigs method.

LocalInventory

The inventory information at a place (e.g. a store) identified by a place ID.

Model

Metadata that describes the training and serving parameters of a [Model][google.cloud.retail.v2.Model]. A [Model][google.cloud.retail.v2.Model] can be associated with a [ServingConfig][google.cloud.retail.v2.ServingConfig] and then queried through the Predict API.

PauseModelRequest

Request for pausing training of a model.

PredictRequest

Request message for Predict method.

PredictResponse

Response message for predict method.

PriceInfo

The price information of a [Product][google.cloud.retail.v2.Product].

Product

Product captures all metadata information of items to be recommended or searched.

ProductDetail

Detailed product information associated with a user event.

ProductInlineSource

The inline source for the input config for ImportProducts method.

ProductInputConfig

The input config source for products.

ProductLevelConfig

Configures what level the product should be uploaded with regards to how users will be send events and how predictions will be made.

Promotion

Promotion information.

PurchaseTransaction

A transaction represents the entire purchase transaction.

PurgeMetadata

Metadata related to the progress of the Purge operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeUserEventsRequest

Request message for PurgeUserEvents method.

PurgeUserEventsResponse

Response of the PurgeUserEventsRequest. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

Rating

The rating of a [Product][google.cloud.retail.v2.Product].

RejoinUserEventsMetadata

Metadata for RejoinUserEvents method.

RejoinUserEventsRequest

Request message for RejoinUserEvents method.

RejoinUserEventsResponse

Response message for RejoinUserEvents method.

RemoveCatalogAttributeRequest

Request for [CatalogService.RemoveCatalogAttribute][google.cloud.retail.v2.CatalogService.RemoveCatalogAttribute] method.

RemoveControlRequest

Request for RemoveControl method.

RemoveFulfillmentPlacesMetadata

Metadata related to the progress of the RemoveFulfillmentPlaces operation. Currently empty because there is no meaningful metadata populated from the [ProductService.RemoveFulfillmentPlaces][google.cloud.retail.v2.ProductService.RemoveFulfillmentPlaces] method.

RemoveFulfillmentPlacesRequest

Request message for [ProductService.RemoveFulfillmentPlaces][google.cloud.retail.v2.ProductService.RemoveFulfillmentPlaces] method.

RemoveFulfillmentPlacesResponse

Response of the RemoveFulfillmentPlacesRequest. Currently empty because there is no meaningful response populated from the [ProductService.RemoveFulfillmentPlaces][google.cloud.retail.v2.ProductService.RemoveFulfillmentPlaces] method.

RemoveLocalInventoriesMetadata

Metadata related to the progress of the RemoveLocalInventories operation. Currently empty because there is no meaningful metadata populated from the [ProductService.RemoveLocalInventories][google.cloud.retail.v2.ProductService.RemoveLocalInventories] method.

RemoveLocalInventoriesRequest

Request message for [ProductService.RemoveLocalInventories][google.cloud.retail.v2.ProductService.RemoveLocalInventories] method.

RemoveLocalInventoriesResponse

Response of the [ProductService.RemoveLocalInventories][google.cloud.retail.v2.ProductService.RemoveLocalInventories] API. Currently empty because there is no meaningful response populated from the [ProductService.RemoveLocalInventories][google.cloud.retail.v2.ProductService.RemoveLocalInventories] method.

ReplaceCatalogAttributeRequest

Request for [CatalogService.ReplaceCatalogAttribute][google.cloud.retail.v2.CatalogService.ReplaceCatalogAttribute] method.

ResumeModelRequest

Request for resuming training of a model.

Rule

A rule is a condition-action pair

SearchRequest

Request message for [SearchService.Search][google.cloud.retail.v2.SearchService.Search] method.

SearchResponse

Response message for [SearchService.Search][google.cloud.retail.v2.SearchService.Search] method.

ServingConfig

Configures metadata that is used to generate serving time results (e.g. search results or recommendation predictions).

SetDefaultBranchRequest

Request message to set a specified branch as new default_branch.

SetInventoryMetadata

Metadata related to the progress of the SetInventory operation. Currently empty because there is no meaningful metadata populated from the [ProductService.SetInventory][google.cloud.retail.v2.ProductService.SetInventory] method.

SetInventoryRequest

Request message for [ProductService.SetInventory][google.cloud.retail.v2.ProductService.SetInventory] method.

SetInventoryResponse

Response of the SetInventoryRequest. Currently empty because there is no meaningful response populated from the [ProductService.SetInventory][google.cloud.retail.v2.ProductService.SetInventory] method.

TuneModelMetadata

Metadata associated with a tune operation.

TuneModelRequest

Request to manually start a tuning process now (instead of waiting for the periodically scheduled tuning to happen).

TuneModelResponse

Response associated with a tune operation.

UpdateAttributesConfigRequest

Request for [CatalogService.UpdateAttributesConfig][google.cloud.retail.v2.CatalogService.UpdateAttributesConfig] method.

UpdateCatalogRequest

Request for [CatalogService.UpdateCatalog][google.cloud.retail.v2.CatalogService.UpdateCatalog] method.

UpdateCompletionConfigRequest

Request for [CatalogService.UpdateCompletionConfig][google.cloud.retail.v2.CatalogService.UpdateCompletionConfig] method.

UpdateControlRequest

Request for UpdateControl method.

UpdateModelRequest

Request for updating an existing model.

UpdateProductRequest

Request message for [ProductService.UpdateProduct][google.cloud.retail.v2.ProductService.UpdateProduct] method.

UpdateServingConfigRequest

Request for UpdateServingConfig method.

UserEvent

UserEvent captures all metadata information Retail API needs to know about how end users interact with customers’ website.

UserEventImportSummary

A summary of import result. The UserEventImportSummary summarizes the import status for user events.

UserEventInlineSource

The inline source for the input config for ImportUserEvents method.

UserEventInputConfig

The input config source for user events.

UserInfo

Information of an end user.

WriteUserEventRequest

Request message for WriteUserEvent method.

Enums

AttributeConfigLevel

At which level we offer configuration for attributes.

RecommendationsFilteringOption

If filtering for recommendations is enabled.

SearchSolutionUseCase

The use case of Cloud Retail Search.

SolutionType

The type of solution.