Struct google_api_proto::google::cloud::retail::v2beta::catalog_service_client::CatalogServiceClient

pub struct CatalogServiceClient<T> { /* private fields */ }

Service for managing catalog configuration.

Implementations

impl<T> CatalogServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> CatalogServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_catalogs( &mut self, request: impl IntoRequest<ListCatalogsRequest> ) -> Result<Response<ListCatalogsResponse>, Status>

Lists all the [Catalog][google.cloud.retail.v2beta.Catalog]s associated with the project.

pub async fn update_catalog( &mut self, request: impl IntoRequest<UpdateCatalogRequest> ) -> Result<Response<Catalog>, Status>

Updates the [Catalog][google.cloud.retail.v2beta.Catalog]s.

pub async fn set_default_branch( &mut self, request: impl IntoRequest<SetDefaultBranchRequest> ) -> Result<Response<()>, Status>

Set a specified branch id as default branch. API methods such as [SearchService.Search][google.cloud.retail.v2beta.SearchService.Search], [ProductService.GetProduct][google.cloud.retail.v2beta.ProductService.GetProduct], [ProductService.ListProducts][google.cloud.retail.v2beta.ProductService.ListProducts] will treat requests using “default_branch” to the actual branch id set as default.

For example, if projects/*/locations/*/catalogs/*/branches/1 is set as default, setting [SearchRequest.branch][google.cloud.retail.v2beta.SearchRequest.branch] to projects/*/locations/*/catalogs/*/branches/default_branch is equivalent to setting [SearchRequest.branch][google.cloud.retail.v2beta.SearchRequest.branch] to projects/*/locations/*/catalogs/*/branches/1.

Using multiple branches can be useful when developers would like to have a staging branch to test and verify for future usage. When it becomes ready, developers switch on the staging branch using this API while keeping using projects/*/locations/*/catalogs/*/branches/default_branch as [SearchRequest.branch][google.cloud.retail.v2beta.SearchRequest.branch] to route the traffic to this staging branch.

CAUTION: If you have live predict/search traffic, switching the default branch could potentially cause outages if the ID space of the new branch is very different from the old one.

More specifically:

PredictionService will only return product IDs from branch {newBranch}.

SearchService will only return product IDs from branch {newBranch} (if branch is not explicitly set).

UserEventService will only join events with products from branch {newBranch}.

pub async fn get_default_branch( &mut self, request: impl IntoRequest<GetDefaultBranchRequest> ) -> Result<Response<GetDefaultBranchResponse>, Status>

Get which branch is currently default branch set by [CatalogService.SetDefaultBranch][google.cloud.retail.v2beta.CatalogService.SetDefaultBranch] method under a specified parent catalog.

pub async fn get_completion_config( &mut self, request: impl IntoRequest<GetCompletionConfigRequest> ) -> Result<Response<CompletionConfig>, Status>

Gets a [CompletionConfig][google.cloud.retail.v2beta.CompletionConfig].

pub async fn update_completion_config( &mut self, request: impl IntoRequest<UpdateCompletionConfigRequest> ) -> Result<Response<CompletionConfig>, Status>

Updates the [CompletionConfig][google.cloud.retail.v2beta.CompletionConfig]s.

pub async fn get_attributes_config( &mut self, request: impl IntoRequest<GetAttributesConfigRequest> ) -> Result<Response<AttributesConfig>, Status>

Gets an [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig].

pub async fn update_attributes_config( &mut self, request: impl IntoRequest<UpdateAttributesConfigRequest> ) -> Result<Response<AttributesConfig>, Status>

Updates the [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig].

The catalog attributes in the request will be updated in the catalog, or inserted if they do not exist. Existing catalog attributes not included in the request will remain unchanged. Attributes that are assigned to products, but do not exist at the catalog level, are always included in the response. The product attribute is assigned default values for missing catalog attribute fields, e.g., searchable and dynamic facetable options.

pub async fn add_catalog_attribute( &mut self, request: impl IntoRequest<AddCatalogAttributeRequest> ) -> Result<Response<AttributesConfig>, Status>

Adds the specified [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] to the [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig].

If the [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] to add already exists, an ALREADY_EXISTS error is returned.

pub async fn remove_catalog_attribute( &mut self, request: impl IntoRequest<RemoveCatalogAttributeRequest> ) -> Result<Response<AttributesConfig>, Status>

Removes the specified [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] from the [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig].

If the [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] to remove does not exist, a NOT_FOUND error is returned.

pub async fn batch_remove_catalog_attributes( &mut self, request: impl IntoRequest<BatchRemoveCatalogAttributesRequest> ) -> Result<Response<BatchRemoveCatalogAttributesResponse>, Status>

Removes all specified [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute]s from the [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig].

pub async fn replace_catalog_attribute( &mut self, request: impl IntoRequest<ReplaceCatalogAttributeRequest> ) -> Result<Response<AttributesConfig>, Status>

Replaces the specified [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] in the [AttributesConfig][google.cloud.retail.v2beta.AttributesConfig] by updating the catalog attribute with the same [CatalogAttribute.key][google.cloud.retail.v2beta.CatalogAttribute.key].

If the [CatalogAttribute][google.cloud.retail.v2beta.CatalogAttribute] to replace does not exist, a NOT_FOUND error is returned.