Module google_api_proto::google::cloud::run::v2

Modules

binary_authorization

Nested message and enum types in BinaryAuthorization.

condition

Nested message and enum types in Condition.

env_var

Nested message and enum types in EnvVar.

revisions_client

Generated client implementations.

services_client

Generated client implementations.

volume

Nested message and enum types in Volume.

vpc_access

Nested message and enum types in VpcAccess.

Structs

BinaryAuthorization

Settings for Binary Authorization feature.

CloudSqlInstance

Represents a specific Cloud SQL instance.

Condition

Defines a status condition for a resource.

Container

A single application container. This specifies both the container to run, the command to run in the container and the arguments to supply to it. Note that additional arguments may be supplied by the system to the container at runtime.

ContainerPort

ContainerPort represents a network port in a single container.

CreateServiceRequest

Request message for creating a Service.

DeleteRevisionRequest

Request message for deleting a retired Revision. Revision lifecycle is usually managed by making changes to the parent Service. Only retired revisions can be deleted with this API.

DeleteServiceRequest

Request message to delete a Service by its full name.

EnvVar

EnvVar represents an environment variable present in a Container.

EnvVarSource

EnvVarSource represents a source for the value of an EnvVar.

GetRevisionRequest

Request message for obtaining a Revision by its full name.

GetServiceRequest

Request message for obtaining a Service by its full name.

ListRevisionsRequest

Request message for retrieving a list of Revisions.

ListRevisionsResponse

Response message containing a list of Revisions.

ListServicesRequest

Request message for retrieving a list of Services.

ListServicesResponse

Response message containing a list of Services.

ResourceRequirements

ResourceRequirements describes the compute resource requirements.

Revision

A Revision is an immutable snapshot of code and configuration. A Revision references a container image. Revisions are only created by updates to its parent Service.

RevisionScaling

Settings for revision-level scaling settings.

RevisionTemplate

RevisionTemplate describes the data a revision should have when created from a template.

SecretKeySelector

SecretEnvVarSource represents a source for the value of an EnvVar.

SecretVolumeSource

The secret’s value will be presented as the content of a file whose name is defined in the item path. If no items are defined, the name of the file is the secret.

Service

Service acts as a top-level container that manages a set of configurations and revision templates which implement a network service. Service exists to provide a singular abstraction which can be access controlled, reasoned about, and which encapsulates software lifecycle decisions such as rollout policy and team resource ownership.

TrafficTarget

Holds a single traffic routing entry for the Service. Allocations can be done to a specific Revision name, or pointing to the latest Ready Revision.

TrafficTargetStatus

Represents the observed state of a single TrafficTarget entry.

UpdateServiceRequest

Request message for updating a service.

VersionToPath

VersionToPath maps a specific version of a secret to a relative file to mount to, relative to VolumeMount’s mount_path.

Volume

Volume represents a named volume in a container.

VolumeMount

VolumeMount describes a mounting of a Volume within a container.

VpcAccess

VPC Access settings. For more information on creating a VPC Connector, visit https://cloud.google.com/vpc/docs/configure-serverless-vpc-access For information on how to configure Cloud Run with an existing VPC Connector, visit https://cloud.google.com/run/docs/configuring/connecting-vpc

Enums

ExecutionEnvironment

Alternatives for execution environments.

IngressTraffic

Allowed ingress traffic for the Container.

TrafficTargetAllocationType

The type of instance allocation.