Module google_api_proto::google::cloud::scheduler::v1

Modules

cloud_scheduler_client

Generated client implementations.

http_target

Nested message and enum types in HttpTarget.

job

Nested message and enum types in Job.

Structs

AppEngineHttpTarget

App Engine target. The job will be pushed to a job handler by means of an HTTP request via an [http_method][google.cloud.scheduler.v1.AppEngineHttpTarget.http_method] such as HTTP POST, HTTP GET, etc. The job is acknowledged by means of an HTTP response code in the range [200 - 299]. Error 503 is considered an App Engine system error instead of an application error. Requests returning error 503 will be retried regardless of retry configuration and not counted against retry counts. Any other response code, or a failure to receive a response before the deadline, constitutes a failed attempt.

AppEngineRouting

App Engine Routing.

CreateJobRequest

Request message for [CreateJob][google.cloud.scheduler.v1.CloudScheduler.CreateJob].

DeleteJobRequest

Request message for deleting a job using [DeleteJob][google.cloud.scheduler.v1.CloudScheduler.DeleteJob].

GetJobRequest

Request message for [GetJob][google.cloud.scheduler.v1.CloudScheduler.GetJob].

HttpTarget

Http target. The job will be pushed to the job handler by means of an HTTP request via an [http_method][google.cloud.scheduler.v1.HttpTarget.http_method] such as HTTP POST, HTTP GET, etc. The job is acknowledged by means of an HTTP response code in the range [200 - 299]. A failure to receive a response constitutes a failed execution. For a redirected request, the response returned by the redirected request is considered.

Job

Configuration for a job. The maximum allowed size for a job is 100KB.

ListJobsRequest

Request message for listing jobs using [ListJobs][google.cloud.scheduler.v1.CloudScheduler.ListJobs].

ListJobsResponse

Response message for listing jobs using [ListJobs][google.cloud.scheduler.v1.CloudScheduler.ListJobs].

OAuthToken

Contains information needed for generating an OAuth token. This type of authorization should generally only be used when calling Google APIs hosted on *.googleapis.com.

OidcToken

Contains information needed for generating an OpenID Connect token. This type of authorization can be used for many scenarios, including calling Cloud Run, or endpoints where you intend to validate the token yourself.

PauseJobRequest

Request message for [PauseJob][google.cloud.scheduler.v1.CloudScheduler.PauseJob].

PubsubTarget

Pub/Sub target. The job will be delivered by publishing a message to the given Pub/Sub topic.

ResumeJobRequest

Request message for [ResumeJob][google.cloud.scheduler.v1.CloudScheduler.ResumeJob].

RetryConfig

Settings that determine the retry behavior.

RunJobRequest

Request message for forcing a job to run now using [RunJob][google.cloud.scheduler.v1.CloudScheduler.RunJob].

UpdateJobRequest

Request message for [UpdateJob][google.cloud.scheduler.v1.CloudScheduler.UpdateJob].

Enums

HttpMethod

The HTTP method used to execute the job.