Module google_api_proto::google::cloud::secretmanager::v1

Modules

replication

Nested message and enum types in Replication.

replication_status

Nested message and enum types in ReplicationStatus.

secret

Nested message and enum types in Secret.

secret_manager_service_client

Generated client implementations.

secret_version

Nested message and enum types in SecretVersion.

Structs

AccessSecretVersionRequest

Request message for [SecretManagerService.AccessSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.AccessSecretVersion].

AccessSecretVersionResponse

Response message for [SecretManagerService.AccessSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.AccessSecretVersion].

AddSecretVersionRequest

Request message for [SecretManagerService.AddSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.AddSecretVersion].

CreateSecretRequest

Request message for [SecretManagerService.CreateSecret][google.cloud.secretmanager.v1.SecretManagerService.CreateSecret].

CustomerManagedEncryption

Configuration for encrypting secret payloads using customer-managed encryption keys (CMEK).

CustomerManagedEncryptionStatus

Describes the status of customer-managed encryption.

DeleteSecretRequest

Request message for [SecretManagerService.DeleteSecret][google.cloud.secretmanager.v1.SecretManagerService.DeleteSecret].

DestroySecretVersionRequest

Request message for [SecretManagerService.DestroySecretVersion][google.cloud.secretmanager.v1.SecretManagerService.DestroySecretVersion].

DisableSecretVersionRequest

Request message for [SecretManagerService.DisableSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.DisableSecretVersion].

EnableSecretVersionRequest

Request message for [SecretManagerService.EnableSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.EnableSecretVersion].

GetSecretRequest

Request message for [SecretManagerService.GetSecret][google.cloud.secretmanager.v1.SecretManagerService.GetSecret].

GetSecretVersionRequest

Request message for [SecretManagerService.GetSecretVersion][google.cloud.secretmanager.v1.SecretManagerService.GetSecretVersion].

ListSecretVersionsRequest

Request message for [SecretManagerService.ListSecretVersions][google.cloud.secretmanager.v1.SecretManagerService.ListSecretVersions].

ListSecretVersionsResponse

Response message for [SecretManagerService.ListSecretVersions][google.cloud.secretmanager.v1.SecretManagerService.ListSecretVersions].

ListSecretsRequest

Request message for [SecretManagerService.ListSecrets][google.cloud.secretmanager.v1.SecretManagerService.ListSecrets].

ListSecretsResponse

Response message for [SecretManagerService.ListSecrets][google.cloud.secretmanager.v1.SecretManagerService.ListSecrets].

Replication

A policy that defines the replication and encryption configuration of data.

ReplicationStatus

The replication status of a [SecretVersion][google.cloud.secretmanager.v1.SecretVersion].

Rotation

The rotation time and period for a [Secret][google.cloud.secretmanager.v1.Secret]. At next_rotation_time, Secret Manager will send a Pub/Sub notification to the topics configured on the Secret. [Secret.topics][google.cloud.secretmanager.v1.Secret.topics] must be set to configure rotation.

Secret

A [Secret][google.cloud.secretmanager.v1.Secret] is a logical secret whose value and versions can be accessed.

SecretPayload

A secret payload resource in the Secret Manager API. This contains the sensitive secret payload that is associated with a [SecretVersion][google.cloud.secretmanager.v1.SecretVersion].

SecretVersion

A secret version resource in the Secret Manager API.

Topic

A Pub/Sub topic which Secret Manager will publish to when control plane events occur on this secret.

UpdateSecretRequest

Request message for [SecretManagerService.UpdateSecret][google.cloud.secretmanager.v1.SecretManagerService.UpdateSecret].