Module google_api_proto::google::cloud::securitycenter::settings::v1beta1

Modules

component_settings

Nested message and enum types in ComponentSettings.

security_center_settings_service_client

Generated client implementations.

security_health_analytics_settings

Nested message and enum types in SecurityHealthAnalyticsSettings.

settings

Nested message and enum types in Settings.

Structs

BatchCalculateEffectiveSettingsRequest

Request message for BatchGetEffectiveSettings.

BatchCalculateEffectiveSettingsResponse

Response message for BatchGetEffectiveSettings.

BatchGetSettingsRequest

Request message for BatchGetSettings.

BatchGetSettingsResponse

Response message for BatchGetSettings.

BillingSettings

Billing settings

CalculateEffectiveComponentSettingsRequest

Request message for CalculateEffectiveComponentSettings.

CalculateEffectiveSettingsRequest

Request message for CalculateEffectiveSettings.

ComponentSettings

Component Settings for Security Command Center

ContainerThreatDetectionSettings

User specified settings for KTD

Detector

Detector is a set of detectors or scanners act as individual checks done within a component e.g. bad IP, bad domains, IAM anomaly, cryptomining, open firewall, etc. Detector is independent of Organization, meaning each detector must be defined for a given Security Center component under a specified billing tier. Organizations can configure the list of detectors based on their subscribed billing tier.

EventThreatDetectionSettings

User specified settings for ETD

GetComponentSettingsRequest

Request message for GetComponentSettings.

GetServiceAccountRequest

Request message for GetServiceAccount.

GetSettingsRequest

Request message for GetSettings.

ListComponentsRequest

Request message for ListComponents.

ListComponentsResponse

Response message for ListComponents.

ListDetectorsRequest

Request message for ListDetectors.

ListDetectorsResponse

Response message for ListDetectors.

ResetComponentSettingsRequest

Request message for ResetComponentSettings.

ResetSettingsRequest

Request message for ResetSettings.

SecurityHealthAnalyticsSettings

User specified settings for Security Health Analytics

ServiceAccount

An organization-level service account to be used by threat detection components.

Settings

Common configuration settings for all of Security Center.

SinkSettings

Sink Settings for Security Command Center

UpdateComponentSettingsRequest

Request message for UpdateComponentSettings.

UpdateSettingsRequest

Request message for UpdateSettings.

WebSecurityScanner

User specified settings for Web Security Scanner

Enums

BillingTier

Billing tier options

BillingType

Billing type

ComponentEnablementState

Valid states for a component