Module google_api_proto::google::cloud::securitycenter::v1

Modules

asset

Nested message and enum types in Asset.

cloud_dlp_data_profile

Nested message and enum types in CloudDlpDataProfile.

connection

Nested message and enum types in Connection.

custom_config

Nested message and enum types in CustomConfig.

cvssv3

Nested message and enum types in Cvssv3.

effective_security_health_analytics_custom_module

Nested message and enum types in EffectiveSecurityHealthAnalyticsCustomModule.

finding

Nested message and enum types in Finding.

iam_binding

Nested message and enum types in IamBinding.

indicator

Nested message and enum types in Indicator.

kubernetes

Nested message and enum types in Kubernetes.

list_assets_response

Nested message and enum types in ListAssetsResponse.

list_findings_response

Nested message and enum types in ListFindingsResponse.

mitre_attack

Nested message and enum types in MitreAttack.

notification_config

Nested message and enum types in NotificationConfig.

notification_message

Nested message and enum types in NotificationMessage.

organization_settings

Nested message and enum types in OrganizationSettings.

run_asset_discovery_response

Nested message and enum types in RunAssetDiscoveryResponse.

security_center_client

Generated client implementations.

security_health_analytics_custom_module

Nested message and enum types in SecurityHealthAnalyticsCustomModule.

simulate_security_health_analytics_custom_module_request

Nested message and enum types in SimulateSecurityHealthAnalyticsCustomModuleRequest.

simulate_security_health_analytics_custom_module_response

Nested message and enum types in SimulateSecurityHealthAnalyticsCustomModuleResponse.

Structs

Access

Represents an access event.

Asset

Security Command Center representation of a Google Cloud resource.

BigQueryExport

Configures how to deliver Findings to BigQuery Instance.

BulkMuteFindingsRequest

Request message for bulk findings update.

BulkMuteFindingsResponse

The response to a BulkMute request. Contains the LRO information.

CloudDlpDataProfile

The data profile associated with the finding.

CloudDlpInspection

Details about the Cloud Data Loss Prevention (Cloud DLP) inspection job that produced the finding.

Compliance

Contains compliance information about a security standard indicating unmet recommendations.

Connection

Contains information about the IP connection associated with the finding.

Contact

The email address of a contact.

ContactDetails

Details about specific contacts

Container

Container associated with the finding.

CreateBigQueryExportRequest

Request message for creating a BigQuery export.

CreateFindingRequest

Request message for creating a finding.

CreateMuteConfigRequest

Request message for creating a mute config.

CreateNotificationConfigRequest

Request message for creating a notification config.

CreateSecurityHealthAnalyticsCustomModuleRequest

Request message for creating Security Health Analytics custom modules.

CreateSourceRequest

Request message for creating a source.

CustomConfig

Defines the properties in a custom module configuration for Security Health Analytics. Use the custom module configuration to create custom detectors that generate custom findings for resources that you specify.

Cve

CVE stands for Common Vulnerabilities and Exposures. More information: https://cve.mitre.org

Cvssv3

Common Vulnerability Scoring System version 3.

Database

Represents database access information, such as queries. A database may be a sub-resource of an instance (as in the case of Cloud SQL instances or Cloud Spanner instances), or the database instance itself. Some database resources might not have the full resource name populated because these resource types, such as Cloud SQL databases, are not yet supported by Cloud Asset Inventory. In these cases only the display name is provided. Some database resources may not have the full resource name populated because these resource types are not yet supported by Cloud Asset Inventory (e.g. Cloud SQL databases). In these cases only the display name will be provided.

DeleteBigQueryExportRequest

Request message for deleting a BigQuery export.

DeleteMuteConfigRequest

Request message for deleting a mute config.

DeleteNotificationConfigRequest

Request message for deleting a notification config.

DeleteSecurityHealthAnalyticsCustomModuleRequest

Request message for deleting Security Health Analytics custom modules.

EffectiveSecurityHealthAnalyticsCustomModule

An EffectiveSecurityHealthAnalyticsCustomModule is the representation of a Security Health Analytics custom module at a specified level of the resource hierarchy: organization, folder, or project. If a custom module is inherited from a parent organization or folder, the value of the enablementState property in EffectiveSecurityHealthAnalyticsCustomModule is set to the value that is effective in the parent, instead of INHERITED. For example, if the module is enabled in a parent organization or folder, the effective enablement_state for the module in all child folders or projects is also enabled. EffectiveSecurityHealthAnalyticsCustomModule is read-only.

EnvironmentVariable

A name-value pair representing an environment variable used in an operating system process.

ExfilResource

Resource where data was exfiltrated from or exfiltrated to.

Exfiltration

Exfiltration represents a data exfiltration attempt from one or more sources to one or more targets. The sources attribute lists the sources of the exfiltrated data. The targets attribute lists the destinations the data was copied to.

ExternalSystem

Representation of third party SIEM/SOAR fields within SCC.

File

File information about the related binary/library used by an executable, or the script used by a script interpreter

Finding

Security Command Center finding.

Folder

Message that contains the resource name and display name of a folder resource.

Geolocation

Represents a geographical location for a given access.

GetBigQueryExportRequest

Request message for retrieving a BigQuery export.

GetEffectiveSecurityHealthAnalyticsCustomModuleRequest

Request message for getting effective Security Health Analytics custom modules.

GetMuteConfigRequest

Request message for retrieving a mute config.

GetNotificationConfigRequest

Request message for getting a notification config.

GetOrganizationSettingsRequest

Request message for getting organization settings.

GetSecurityHealthAnalyticsCustomModuleRequest

Request message for getting Security Health Analytics custom modules.

GetSourceRequest

Request message for getting a source.

GroupAssetsRequest

Request message for grouping by assets.

GroupAssetsResponse

Response message for grouping by assets.

GroupFindingsRequest

Request message for grouping by findings.

GroupFindingsResponse

Response message for group by findings.

GroupResult

Result containing the properties and count of a groupBy request.

IamBinding

Represents a particular IAM binding, which captures a member’s role addition, removal, or state.

Indicator

Represents what’s commonly known as an indicator of compromise (IoC) in computer forensics. This is an artifact observed on a network or in an operating system that, with high confidence, indicates a computer intrusion. For more information, see Indicator of compromise.

KernelRootkit

Kernel mode rootkit signatures.

Kubernetes

Kubernetes-related attributes.

Label

Represents a generic name-value label. A label has separate name and value fields to support filtering with the contains() function. For more information, see Filtering on array-type fields.

ListAssetsRequest

Request message for listing assets.

ListAssetsResponse

Response message for listing assets.

ListBigQueryExportsRequest

Request message for listing BigQuery exports at a given scope e.g. organization, folder or project.

ListBigQueryExportsResponse

Response message for listing BigQuery exports.

ListDescendantSecurityHealthAnalyticsCustomModulesRequest

Request message for listing descendant Security Health Analytics custom modules.

ListDescendantSecurityHealthAnalyticsCustomModulesResponse

Response message for listing descendant Security Health Analytics custom modules.

ListEffectiveSecurityHealthAnalyticsCustomModulesRequest

Request message for listing effective Security Health Analytics custom modules.

ListEffectiveSecurityHealthAnalyticsCustomModulesResponse

Response message for listing effective Security Health Analytics custom modules.

ListFindingsRequest

Request message for listing findings.

ListFindingsResponse

Response message for listing findings.

ListMuteConfigsRequest

Request message for listing mute configs at a given scope e.g. organization, folder or project.

ListMuteConfigsResponse

Response message for listing mute configs.

ListNotificationConfigsRequest

Request message for listing notification configs.

ListNotificationConfigsResponse

Response message for listing notification configs.

ListSecurityHealthAnalyticsCustomModulesRequest

Request message for listing Security Health Analytics custom modules.

ListSecurityHealthAnalyticsCustomModulesResponse

Response message for listing Security Health Analytics custom modules.

ListSourcesRequest

Request message for listing sources.

ListSourcesResponse

Response message for listing sources.

MitreAttack

MITRE ATT&CK tactics and techniques related to this finding. See: https://attack.mitre.org

MuteConfig

A mute config is a Cloud SCC resource that contains the configuration to mute create/update events of findings.

NotificationConfig

Cloud Security Command Center (Cloud SCC) notification configs.

NotificationMessage

Cloud SCC’s Notification

OrganizationSettings

User specified settings that are attached to the Security Command Center organization.

Process

Represents an operating system process.

Reference

Additional Links

Resource

Information related to the Google Cloud resource.

RunAssetDiscoveryRequest

Request message for running asset discovery for an organization.

RunAssetDiscoveryResponse

Response of asset discovery run

SecurityHealthAnalyticsCustomModule

Represents an instance of a Security Health Analytics custom module, including its full module name, display name, enablement state, and last updated time. You can create a custom module at the organization, folder, or project level. Custom modules that you create at the organization or folder level are inherited by the child folders and projects.

SecurityMarks

User specified security marks that are attached to the parent Security Command Center resource. Security marks are scoped within a Security Command Center organization – they can be modified and viewed by all users who have proper permissions on the organization.

ServiceAccountDelegationInfo

Identity delegation history of an authenticated service account.

SetFindingStateRequest

Request message for updating a finding’s state.

SetMuteRequest

Request message for updating a finding’s mute status.

SimulateSecurityHealthAnalyticsCustomModuleRequest

Request message to simulate a CustomConfig against a given test resource. Maximum size of the request is 4 MB by default.

SimulateSecurityHealthAnalyticsCustomModuleResponse

Response message for simulating a SecurityHealthAnalyticsCustomModule against a given resource.

Source

Security Command Center finding source. A finding source is an entity or a mechanism that can produce a finding. A source is like a container of findings that come from the same scanner, logger, monitor, and other tools.

UpdateBigQueryExportRequest

Request message for updating a BigQuery export.

UpdateExternalSystemRequest

Request message for updating a ExternalSystem resource.

UpdateFindingRequest

Request message for updating or creating a finding.

UpdateMuteConfigRequest

Request message for updating a mute config.

UpdateNotificationConfigRequest

Request message for updating a notification config.

UpdateOrganizationSettingsRequest

Request message for updating an organization’s settings.

UpdateSecurityHealthAnalyticsCustomModuleRequest

Request message for updating Security Health Analytics custom modules.

UpdateSecurityMarksRequest

Request message for updating a SecurityMarks resource.

UpdateSourceRequest

Request message for updating a source.

Vulnerability

Refers to common vulnerability fields e.g. cve, cvss, cwe etc.