Module google_api_proto::google::cloud::securitycenter::v1

Modules

asset

Nested message and enum types in Asset.

cvssv3

Nested message and enum types in Cvssv3.

finding

Nested message and enum types in Finding.

iam_binding

Nested message and enum types in IamBinding.

list_assets_response

Nested message and enum types in ListAssetsResponse.

list_findings_response

Nested message and enum types in ListFindingsResponse.

mitre_attack

Nested message and enum types in MitreAttack.

notification_config

Nested message and enum types in NotificationConfig.

notification_message

Nested message and enum types in NotificationMessage.

organization_settings

Nested message and enum types in OrganizationSettings.

run_asset_discovery_response

Nested message and enum types in RunAssetDiscoveryResponse.

security_center_client

Generated client implementations.

Structs

Access

Represents an access event.

Asset

Security Command Center representation of a Google Cloud resource.

BigQueryExport

Configures how to deliver Findings to BigQuery Instance.

BulkMuteFindingsRequest

Request message for bulk findings update.

BulkMuteFindingsResponse

The response to a BulkMute request. Contains the LRO information.

CreateBigQueryExportRequest

Request message for creating a big query export.

CreateFindingRequest

Request message for creating a finding.

CreateMuteConfigRequest

Request message for creating a mute config.

CreateNotificationConfigRequest

Request message for creating a notification config.

CreateSourceRequest

Request message for creating a source.

Cve

CVE stands for Common Vulnerabilities and Exposures. More information: https://cve.mitre.org

Cvssv3

Common Vulnerability Scoring System version 3.

DeleteBigQueryExportRequest

Request message for deleting a big query export.

DeleteMuteConfigRequest

Request message for deleting a mute config.

DeleteNotificationConfigRequest

Request message for deleting a notification config.

ExternalSystem

Representation of third party SIEM/SOAR fields within SCC.

Finding

Security Command Center finding.

Folder

Message that contains the resource name and display name of a folder resource.

Geolocation

Represents a geographical location for a given access.

GetBigQueryExportRequest

Request message for retrieving a big query export.

GetMuteConfigRequest

Request message for retrieving a mute config.

GetNotificationConfigRequest

Request message for getting a notification config.

GetOrganizationSettingsRequest

Request message for getting organization settings.

GetSourceRequest

Request message for getting a source.

GroupAssetsRequest

Request message for grouping by assets.

GroupAssetsResponse

Response message for grouping by assets.

GroupFindingsRequest

Request message for grouping by findings.

GroupFindingsResponse

Response message for group by findings.

GroupResult

Result containing the properties and count of a groupBy request.

IamBinding

Represents a particular IAM binding, which captures a member’s role addition, removal, or state.

Indicator

Represents what’s commonly known as an Indicator of compromise (IoC) in computer forensics. This is an artifact observed on a network or in an operating system that, with high confidence, indicates a computer intrusion. Reference: https://en.wikipedia.org/wiki/Indicator_of_compromise

ListAssetsRequest

Request message for listing assets.

ListAssetsResponse

Response message for listing assets.

ListBigQueryExportsRequest

Request message for listing BigQuery exports at a given scope e.g. organization, folder or project.

ListBigQueryExportsResponse

Response message for listing BigQuery exports.

ListFindingsRequest

Request message for listing findings.

ListFindingsResponse

Response message for listing findings.

ListMuteConfigsRequest

Request message for listing mute configs at a given scope e.g. organization, folder or project.

ListMuteConfigsResponse

Response message for listing mute configs.

ListNotificationConfigsRequest

Request message for listing notification configs.

ListNotificationConfigsResponse

Response message for listing notification configs.

ListSourcesRequest

Request message for listing sources.

ListSourcesResponse

Response message for listing sources.

MitreAttack

MITRE ATT&CK tactics and techniques related to this finding. See: https://attack.mitre.org

MuteConfig

A mute config is a Cloud SCC resource that contains the configuration to mute create/update events of findings.

NotificationConfig

Cloud Security Command Center (Cloud SCC) notification configs.

NotificationMessage

Cloud SCC’s Notification

OrganizationSettings

User specified settings that are attached to the Security Command Center organization.

Reference

Additional Links

Resource

Information related to the Google Cloud resource.

RunAssetDiscoveryRequest

Request message for running asset discovery for an organization.

RunAssetDiscoveryResponse

Response of asset discovery run

SecurityMarks

User specified security marks that are attached to the parent Security Command Center resource. Security marks are scoped within a Security Command Center organization – they can be modified and viewed by all users who have proper permissions on the organization.

SetFindingStateRequest

Request message for updating a finding’s state.

SetMuteRequest

Request message for updating a finding’s mute status.

Source

Security Command Center finding source. A finding source is an entity or a mechanism that can produce a finding. A source is like a container of findings that come from the same scanner, logger, monitor, and other tools.

UpdateBigQueryExportRequest

Request message for updating a BigQuery export.

UpdateExternalSystemRequest

Request message for updating a ExternalSystem resource.

UpdateFindingRequest

Request message for updating or creating a finding.

UpdateMuteConfigRequest

Request message for updating a mute config.

UpdateNotificationConfigRequest

Request message for updating a notification config.

UpdateOrganizationSettingsRequest

Request message for updating an organization’s settings.

UpdateSecurityMarksRequest

Request message for updating a SecurityMarks resource.

UpdateSourceRequest

Request message for updating a source.

Vulnerability

Refers to common vulnerability fields e.g. cve, cvss, cwe etc.