Module google_api_proto::google::cloud::securitycentermanagement::v1

Modules

custom_config

Nested message and enum types in CustomConfig.

effective_event_threat_detection_custom_module

Nested message and enum types in EffectiveEventThreatDetectionCustomModule.

effective_security_health_analytics_custom_module

Nested message and enum types in EffectiveSecurityHealthAnalyticsCustomModule.

event_threat_detection_custom_module

Nested message and enum types in EventThreatDetectionCustomModule.

security_center_management_client

Generated client implementations.

security_center_service

Nested message and enum types in SecurityCenterService.

security_health_analytics_custom_module

Nested message and enum types in SecurityHealthAnalyticsCustomModule.

simulate_security_health_analytics_custom_module_request

Nested message and enum types in SimulateSecurityHealthAnalyticsCustomModuleRequest.

simulate_security_health_analytics_custom_module_response

Nested message and enum types in SimulateSecurityHealthAnalyticsCustomModuleResponse.

simulated_finding

Nested message and enum types in SimulatedFinding.

validate_event_threat_detection_custom_module_response

Nested message and enum types in ValidateEventThreatDetectionCustomModuleResponse.

Structs

CreateEventThreatDetectionCustomModuleRequest

Message for creating a EventThreatDetectionCustomModule

CreateSecurityHealthAnalyticsCustomModuleRequest

Message for creating a SecurityHealthAnalyticsCustomModule

CustomConfig

Defines the properties in a custom module configuration for Security Health Analytics. Use the custom module configuration to create custom detectors that generate custom findings for resources that you specify.

DeleteEventThreatDetectionCustomModuleRequest

Message for deleting a EventThreatDetectionCustomModule

DeleteSecurityHealthAnalyticsCustomModuleRequest

Message for deleting a SecurityHealthAnalyticsCustomModule

EffectiveEventThreatDetectionCustomModule

An EffectiveEventThreatDetectionCustomModule is the representation of EventThreatDetectionCustomModule at a given level taking hierarchy into account and resolving various fields accordingly. e.g. if the module is enabled at the ancestor level, effective modules at all descendant levels will have enablement_state set to ENABLED. Similarly, if module.inherited is set, then effective module’s config will contain the ancestor’s config details. EffectiveEventThreatDetectionCustomModule is read-only.

EffectiveSecurityHealthAnalyticsCustomModule

An EffectiveSecurityHealthAnalyticsCustomModule is the representation of a Security Health Analytics custom module at a specified level of the resource hierarchy: organization, folder, or project. If a custom module is inherited from a parent organization or folder, the value of the enablementState property in EffectiveSecurityHealthAnalyticsCustomModule is set to the value that is effective in the parent, instead of INHERITED. For example, if the module is enabled in a parent organization or folder, the effective enablement_state for the module in all child folders or projects is also enabled. EffectiveSecurityHealthAnalyticsCustomModule is read-only.

EventThreatDetectionCustomModule

An event threat detection custom module is a Cloud SCC resource that contains the configuration and enablement state of a custom module, which enables ETD to write certain findings to Cloud SCC.

GetEffectiveEventThreatDetectionCustomModuleRequest

Message for getting a EffectiveEventThreatDetectionCustomModule

GetEffectiveSecurityHealthAnalyticsCustomModuleRequest

Message for getting a EffectiveSecurityHealthAnalyticsCustomModule

GetEventThreatDetectionCustomModuleRequest

Message for getting a EventThreatDetectionCustomModule

GetSecurityCenterServiceRequest

Request message for getting a Security Command Center service.

GetSecurityHealthAnalyticsCustomModuleRequest

Message for getting a SecurityHealthAnalyticsCustomModule

ListDescendantEventThreatDetectionCustomModulesRequest

Request message for listing descendant Event Threat Detection custom modules.

ListDescendantEventThreatDetectionCustomModulesResponse

Response message for listing descendant Event Threat Detection custom modules.

ListDescendantSecurityHealthAnalyticsCustomModulesRequest

Request message for listing descendant Security Health Analytics custom modules.

ListDescendantSecurityHealthAnalyticsCustomModulesResponse

Response message for listing descendant Security Health Analytics custom modules.

ListEffectiveEventThreatDetectionCustomModulesRequest

Request message for listing effective Event Threat Detection custom modules.

ListEffectiveEventThreatDetectionCustomModulesResponse

Response message for listing effective Event Threat Detection custom modules.

ListEffectiveSecurityHealthAnalyticsCustomModulesRequest

Request message for listing effective Security Health Analytics custom modules.

ListEffectiveSecurityHealthAnalyticsCustomModulesResponse

Response message for listing effective Security Health Analytics custom modules.

ListEventThreatDetectionCustomModulesRequest

Request message for listing Event Threat Detection custom modules.

ListEventThreatDetectionCustomModulesResponse

Response message for listing Event Threat Detection custom modules.

ListSecurityCenterServicesRequest

Request message for listing Security Command Center services.

ListSecurityCenterServicesResponse

Response message for listing Security Command Center services.

ListSecurityHealthAnalyticsCustomModulesRequest

Request message for listing Security Health Analytics custom modules.

ListSecurityHealthAnalyticsCustomModulesResponse

Response message for listing Security Health Analytics custom modules.

SecurityCenterService

Represents a particular Security Command Center service. This includes settings information such as top-level enablement in addition to individual module settings. Service settings can be configured at the organization, folder, or project level. Service settings at the organization or folder level are inherited by those in child folders and projects.

SecurityHealthAnalyticsCustomModule

Represents an instance of a Security Health Analytics custom module, including its full module name, display name, enablement state, and last updated time. You can create a custom module at the organization, folder, or project level. Custom modules that you create at the organization or folder level are inherited by the child folders and projects.

SimulateSecurityHealthAnalyticsCustomModuleRequest

Request message to simulate a CustomConfig against a given test resource. Maximum size of the request is 4 MB by default.

SimulateSecurityHealthAnalyticsCustomModuleResponse

Response message for simulating a SecurityHealthAnalyticsCustomModule against a given resource.

SimulatedFinding

A subset of the fields of the Security Center Finding proto. The minimum set of fields needed to represent a simulated finding from a SHA custom module.

UpdateEventThreatDetectionCustomModuleRequest

Message for updating a EventThreatDetectionCustomModule

UpdateSecurityCenterServiceRequest

Request message for updating a Security Command Center service.

UpdateSecurityHealthAnalyticsCustomModuleRequest

Message for updating a SecurityHealthAnalyticsCustomModule

ValidateEventThreatDetectionCustomModuleRequest

Request to validate an Event Threat Detection custom module.

ValidateEventThreatDetectionCustomModuleResponse

Response to validating an Event Threat Detection custom module.