Module google_api_proto::google::cloud::securityposture::v1

Modules

constraint

Nested message and enum types in Constraint.

custom_config

Nested message and enum types in CustomConfig.

custom_constraint

Nested message and enum types in CustomConstraint.

policy

Nested message and enum types in Policy.

policy_rule

Nested message and enum types in PolicyRule.

posture

Nested message and enum types in Posture.

posture_deployment

Nested message and enum types in PostureDeployment.

posture_template

Nested message and enum types in PostureTemplate.

security_posture_client

Generated client implementations.

Structs

Constraint

Representation of a Constraint.

CreatePostureDeploymentRequest

Message for creating a PostureDeployment.

CreatePostureRequest

Message for creating a Posture.

CustomConfig

Defines the properties in a custom module configuration for Security Health Analytics. Use the custom module configuration to create custom detectors that generate custom findings for resources that you specify.

CustomConstraint

A custom constraint defined by customers which can only be applied to the given resource types and organization.

DeletePostureDeploymentRequest

Message for deleting a PostureDeployment.

DeletePostureRequest

Message for deleting a Posture.

ExtractPostureRequest

Message for extracting existing policies on a workload as a Posture.

GetPostureDeploymentRequest

Message for getting a PostureDeployment.

GetPostureRequest

Message for getting a Posture.

GetPostureTemplateRequest

Message for getting a Posture Template.

ListPostureDeploymentsRequest

Message for requesting list of PostureDeployments.

ListPostureDeploymentsResponse

Message for response to listing PostureDeployments.

ListPostureRevisionsRequest

Message for requesting list of Posture revisions.

ListPostureRevisionsResponse

Message for response to listing PostureRevisions.

ListPostureTemplatesRequest

Message for requesting list of Posture Templates.

ListPostureTemplatesResponse

Message for response to listing PostureTemplates.

ListPosturesRequest

Message for requesting list of Postures.

ListPosturesResponse

Message for response to listing Postures.

OperationMetadata

Represents the metadata of the long-running operation.

OrgPolicyConstraint

Message for Org Policy Canned Constraint.

OrgPolicyConstraintCustom

Message for Org Policy Custom Constraint.

Policy

Policy representation.

PolicyRule

A rule used to express this policy.

PolicySet

PolicySet representation.

Posture

Postures Definition of a Posture.

PostureDeployment

========================== PostureDeployments ========================== Message describing PostureDeployment resource.

PostureTemplate

PostureTemplates Message describing PostureTemplate object.

SecurityHealthAnalyticsCustomModule

Message for SHA Custom Module

SecurityHealthAnalyticsModule

Message for Security Health Analytics built-in detector.

UpdatePostureDeploymentRequest

Message for updating a PostureDeployment.

UpdatePostureRequest

Message for updating a Posture.

Enums

EnablementState

Possible enablement states of a service or module.