Module google_api_proto::google::cloud::speech::v2

Modules

batch_recognize_file_metadata

Nested message and enum types in BatchRecognizeFileMetadata.

batch_recognize_file_result

Nested message and enum types in BatchRecognizeFileResult.

batch_recognize_request

Nested message and enum types in BatchRecognizeRequest.

custom_class

Nested message and enum types in CustomClass.

explicit_decoding_config

Nested message and enum types in ExplicitDecodingConfig.

operation_metadata

Nested message and enum types in OperationMetadata.

phrase_set

Nested message and enum types in PhraseSet.

recognition_config

Nested message and enum types in RecognitionConfig.

recognition_features

Nested message and enum types in RecognitionFeatures.

recognition_output_config

Nested message and enum types in RecognitionOutputConfig.

recognize_request

Nested message and enum types in RecognizeRequest.

recognizer

Nested message and enum types in Recognizer.

speech_adaptation

Nested message and enum types in SpeechAdaptation.

speech_client

Generated client implementations.

streaming_recognition_features

Nested message and enum types in StreamingRecognitionFeatures.

streaming_recognize_request

Nested message and enum types in StreamingRecognizeRequest.

streaming_recognize_response

Nested message and enum types in StreamingRecognizeResponse.

transcript_normalization

Nested message and enum types in TranscriptNormalization.

Structs

AutoDetectDecodingConfig

Automatically detected decoding parameters. Supported for the following encodings:

BatchRecognizeFileMetadata

Metadata about a single file in a batch for BatchRecognize.

BatchRecognizeFileResult

Final results for a single file.

BatchRecognizeMetadata

Operation metadata for [BatchRecognize][google.cloud.speech.v2.Speech.BatchRecognize].

BatchRecognizeRequest

Request message for the [BatchRecognize][google.cloud.speech.v2.Speech.BatchRecognize] method.

BatchRecognizeResponse

Response message for [BatchRecognize][google.cloud.speech.v2.Speech.BatchRecognize] that is packaged into a longrunning [Operation][google.longrunning.Operation].

BatchRecognizeResults

Output type for Cloud Storage of BatchRecognize transcripts. Though this proto isn’t returned in this API anywhere, the Cloud Storage transcripts will be this proto serialized and should be parsed as such.

BatchRecognizeTranscriptionMetadata

Metadata about transcription for a single file (for example, progress percent).

CloudStorageResult

Final results written to Cloud Storage.

Config

Message representing the config for the Speech-to-Text API. This includes an optional KMS key with which incoming data will be encrypted.

CreateCustomClassRequest

Request message for the [CreateCustomClass][google.cloud.speech.v2.Speech.CreateCustomClass] method.

CreatePhraseSetRequest

Request message for the [CreatePhraseSet][google.cloud.speech.v2.Speech.CreatePhraseSet] method.

CreateRecognizerRequest

Request message for the [CreateRecognizer][google.cloud.speech.v2.Speech.CreateRecognizer] method.

CustomClass

CustomClass for biasing in speech recognition. Used to define a set of words or phrases that represents a common concept or theme likely to appear in your audio, for example a list of passenger ship names.

DeleteCustomClassRequest

Request message for the [DeleteCustomClass][google.cloud.speech.v2.Speech.DeleteCustomClass] method.

DeletePhraseSetRequest

Request message for the [DeletePhraseSet][google.cloud.speech.v2.Speech.DeletePhraseSet] method.

DeleteRecognizerRequest

Request message for the [DeleteRecognizer][google.cloud.speech.v2.Speech.DeleteRecognizer] method.

ExplicitDecodingConfig

Explicitly specified decoding parameters.

GcsOutputConfig

Output configurations for Cloud Storage.

GetConfigRequest

Request message for the [GetConfig][google.cloud.speech.v2.Speech.GetConfig] method.

GetCustomClassRequest

Request message for the [GetCustomClass][google.cloud.speech.v2.Speech.GetCustomClass] method.

GetPhraseSetRequest

Request message for the [GetPhraseSet][google.cloud.speech.v2.Speech.GetPhraseSet] method.

GetRecognizerRequest

Request message for the [GetRecognizer][google.cloud.speech.v2.Speech.GetRecognizer] method.

InlineOutputConfig

Output configurations for inline response.

InlineResult

Final results returned inline in the recognition response.

ListCustomClassesRequest

Request message for the [ListCustomClasses][google.cloud.speech.v2.Speech.ListCustomClasses] method.

ListCustomClassesResponse

Response message for the [ListCustomClasses][google.cloud.speech.v2.Speech.ListCustomClasses] method.

ListPhraseSetsRequest

Request message for the [ListPhraseSets][google.cloud.speech.v2.Speech.ListPhraseSets] method.

ListPhraseSetsResponse

Response message for the [ListPhraseSets][google.cloud.speech.v2.Speech.ListPhraseSets] method.

ListRecognizersRequest

Request message for the [ListRecognizers][google.cloud.speech.v2.Speech.ListRecognizers] method.

ListRecognizersResponse

Response message for the [ListRecognizers][google.cloud.speech.v2.Speech.ListRecognizers] method.

OperationMetadata

Represents the metadata of a long-running operation.

PhraseSet

PhraseSet for biasing in speech recognition. A PhraseSet is used to provide “hints” to the speech recognizer to favor specific words and phrases in the results.

RecognitionConfig

Provides information to the Recognizer that specifies how to process the recognition request.

RecognitionFeatures

Available recognition features.

RecognitionOutputConfig

Configuration options for the output(s) of recognition.

RecognitionResponseMetadata

Metadata about the recognition request and response.

RecognizeRequest

Request message for the [Recognize][google.cloud.speech.v2.Speech.Recognize] method. Either content or uri must be supplied. Supplying both or neither returns [INVALID_ARGUMENT][google.rpc.Code.INVALID_ARGUMENT]. See content limits.

RecognizeResponse

Response message for the [Recognize][google.cloud.speech.v2.Speech.Recognize] method.

Recognizer

A Recognizer message. Stores recognition configuration and metadata.

SpeakerDiarizationConfig

Configuration to enable speaker diarization.

SpeechAdaptation

Provides “hints” to the speech recognizer to favor specific words and phrases in the results. PhraseSets can be specified as an inline resource, or a reference to an existing PhraseSet resource.

SpeechRecognitionAlternative

Alternative hypotheses (a.k.a. n-best list).

SpeechRecognitionResult

A speech recognition result corresponding to a portion of the audio.

StreamingRecognitionConfig

Provides configuration information for the StreamingRecognize request.

StreamingRecognitionFeatures

Available recognition features specific to streaming recognition requests.

StreamingRecognitionResult

A streaming speech recognition result corresponding to a portion of the audio that is currently being processed.

StreamingRecognizeRequest

Request message for the [StreamingRecognize][google.cloud.speech.v2.Speech.StreamingRecognize] method. Multiple [StreamingRecognizeRequest][google.cloud.speech.v2.StreamingRecognizeRequest] messages are sent in one call.

StreamingRecognizeResponse

StreamingRecognizeResponse is the only message returned to the client by StreamingRecognize. A series of zero or more StreamingRecognizeResponse messages are streamed back to the client. If there is no recognizable audio then no messages are streamed back to the client.

TranscriptNormalization

Transcription normalization configuration. Use transcription normalization to automatically replace parts of the transcript with phrases of your choosing. For StreamingRecognize, this normalization only applies to stable partial transcripts (stability > 0.8) and final transcripts.

UndeleteCustomClassRequest

Request message for the [UndeleteCustomClass][google.cloud.speech.v2.Speech.UndeleteCustomClass] method.

UndeletePhraseSetRequest

Request message for the [UndeletePhraseSet][google.cloud.speech.v2.Speech.UndeletePhraseSet] method.

UndeleteRecognizerRequest

Request message for the [UndeleteRecognizer][google.cloud.speech.v2.Speech.UndeleteRecognizer] method.

UpdateConfigRequest

Request message for the [UpdateConfig][google.cloud.speech.v2.Speech.UpdateConfig] method.

UpdateCustomClassRequest

Request message for the [UpdateCustomClass][google.cloud.speech.v2.Speech.UpdateCustomClass] method.

UpdatePhraseSetRequest

Request message for the [UpdatePhraseSet][google.cloud.speech.v2.Speech.UpdatePhraseSet] method.

UpdateRecognizerRequest

Request message for the [UpdateRecognizer][google.cloud.speech.v2.Speech.UpdateRecognizer] method.

WordInfo

Word-specific information for recognized words.