Struct google_api_proto::google::cloud::sql::v1::sql_instances_service_client::SqlInstancesServiceClient

pub struct SqlInstancesServiceClient<T> { /* private fields */ }

Service to manage Cloud SQL instances.

Implementations

impl<T> SqlInstancesServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> SqlInstancesServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn add_server_ca( &mut self, request: impl IntoRequest<SqlInstancesAddServerCaRequest> ) -> Result<Response<Operation>, Status>

Adds a new trusted Certificate Authority (CA) version for the specified instance. Required to prepare for a certificate rotation. If a CA version was previously added but never used in a certificate rotation, this operation replaces that version. There cannot be more than one CA version waiting to be rotated in.

pub async fn clone( &mut self, request: impl IntoRequest<SqlInstancesCloneRequest> ) -> Result<Response<Operation>, Status>

Creates a Cloud SQL instance as a clone of the source instance. Using this operation might cause your instance to restart.

pub async fn delete( &mut self, request: impl IntoRequest<SqlInstancesDeleteRequest> ) -> Result<Response<Operation>, Status>

Deletes a Cloud SQL instance.

pub async fn demote_master( &mut self, request: impl IntoRequest<SqlInstancesDemoteMasterRequest> ) -> Result<Response<Operation>, Status>

Demotes the stand-alone instance to be a Cloud SQL read replica for an external database server.

pub async fn demote( &mut self, request: impl IntoRequest<SqlInstancesDemoteRequest> ) -> Result<Response<Operation>, Status>

Demotes an existing standalone instance to be a Cloud SQL read replica for an external database server.

pub async fn export( &mut self, request: impl IntoRequest<SqlInstancesExportRequest> ) -> Result<Response<Operation>, Status>

Exports data from a Cloud SQL instance to a Cloud Storage bucket as a SQL dump or CSV file.

pub async fn failover( &mut self, request: impl IntoRequest<SqlInstancesFailoverRequest> ) -> Result<Response<Operation>, Status>

Initiates a manual failover of a high availability (HA) primary instance to a standby instance, which becomes the primary instance. Users are then rerouted to the new primary. For more information, see the Overview of high availability page in the Cloud SQL documentation. If using Legacy HA (MySQL only), this causes the instance to failover to its failover replica instance.

pub async fn reencrypt( &mut self, request: impl IntoRequest<SqlInstancesReencryptRequest> ) -> Result<Response<Operation>, Status>

Reencrypt CMEK instance with latest key version.

pub async fn get( &mut self, request: impl IntoRequest<SqlInstancesGetRequest> ) -> Result<Response<DatabaseInstance>, Status>

Retrieves a resource containing information about a Cloud SQL instance.

pub async fn import( &mut self, request: impl IntoRequest<SqlInstancesImportRequest> ) -> Result<Response<Operation>, Status>

Imports data into a Cloud SQL instance from a SQL dump or CSV file in Cloud Storage.

pub async fn insert( &mut self, request: impl IntoRequest<SqlInstancesInsertRequest> ) -> Result<Response<Operation>, Status>

Creates a new Cloud SQL instance.

pub async fn list( &mut self, request: impl IntoRequest<SqlInstancesListRequest> ) -> Result<Response<InstancesListResponse>, Status>

Lists instances under a given project.

pub async fn list_server_cas( &mut self, request: impl IntoRequest<SqlInstancesListServerCasRequest> ) -> Result<Response<InstancesListServerCasResponse>, Status>

Lists all of the trusted Certificate Authorities (CAs) for the specified instance. There can be up to three CAs listed: the CA that was used to sign the certificate that is currently in use, a CA that has been added but not yet used to sign a certificate, and a CA used to sign a certificate that has previously rotated out.

pub async fn patch( &mut self, request: impl IntoRequest<SqlInstancesPatchRequest> ) -> Result<Response<Operation>, Status>

Partially updates settings of a Cloud SQL instance by merging the request with the current configuration. This method supports patch semantics.

pub async fn promote_replica( &mut self, request: impl IntoRequest<SqlInstancesPromoteReplicaRequest> ) -> Result<Response<Operation>, Status>

Promotes the read replica instance to be a stand-alone Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn switchover( &mut self, request: impl IntoRequest<SqlInstancesSwitchoverRequest> ) -> Result<Response<Operation>, Status>

Switches over from the primary instance to the replica instance.

pub async fn reset_ssl_config( &mut self, request: impl IntoRequest<SqlInstancesResetSslConfigRequest> ) -> Result<Response<Operation>, Status>

Deletes all client certificates and generates a new server SSL certificate for the instance.

pub async fn restart( &mut self, request: impl IntoRequest<SqlInstancesRestartRequest> ) -> Result<Response<Operation>, Status>

Restarts a Cloud SQL instance.

pub async fn restore_backup( &mut self, request: impl IntoRequest<SqlInstancesRestoreBackupRequest> ) -> Result<Response<Operation>, Status>

Restores a backup of a Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn rotate_server_ca( &mut self, request: impl IntoRequest<SqlInstancesRotateServerCaRequest> ) -> Result<Response<Operation>, Status>

Rotates the server certificate to one signed by the Certificate Authority (CA) version previously added with the addServerCA method.

pub async fn start_replica( &mut self, request: impl IntoRequest<SqlInstancesStartReplicaRequest> ) -> Result<Response<Operation>, Status>

Starts the replication in the read replica instance.

pub async fn stop_replica( &mut self, request: impl IntoRequest<SqlInstancesStopReplicaRequest> ) -> Result<Response<Operation>, Status>

Stops the replication in the read replica instance.

pub async fn truncate_log( &mut self, request: impl IntoRequest<SqlInstancesTruncateLogRequest> ) -> Result<Response<Operation>, Status>

Truncate MySQL general and slow query log tables MySQL only.

pub async fn update( &mut self, request: impl IntoRequest<SqlInstancesUpdateRequest> ) -> Result<Response<Operation>, Status>

Updates settings of a Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn create_ephemeral( &mut self, request: impl IntoRequest<SqlInstancesCreateEphemeralCertRequest> ) -> Result<Response<SslCert>, Status>

Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.

pub async fn reschedule_maintenance( &mut self, request: impl IntoRequest<SqlInstancesRescheduleMaintenanceRequest> ) -> Result<Response<Operation>, Status>

Reschedules the maintenance on the given instance.

pub async fn verify_external_sync_settings( &mut self, request: impl IntoRequest<SqlInstancesVerifyExternalSyncSettingsRequest> ) -> Result<Response<SqlInstancesVerifyExternalSyncSettingsResponse>, Status>

Verify External primary instance external sync settings.

pub async fn start_external_sync( &mut self, request: impl IntoRequest<SqlInstancesStartExternalSyncRequest> ) -> Result<Response<Operation>, Status>

Start External primary instance migration.

pub async fn perform_disk_shrink( &mut self, request: impl IntoRequest<SqlInstancesPerformDiskShrinkRequest> ) -> Result<Response<Operation>, Status>

Perform Disk Shrink on primary instance.

pub async fn get_disk_shrink_config( &mut self, request: impl IntoRequest<SqlInstancesGetDiskShrinkConfigRequest> ) -> Result<Response<SqlInstancesGetDiskShrinkConfigResponse>, Status>

Get Disk Shrink Config for a given instance.

pub async fn reset_replica_size( &mut self, request: impl IntoRequest<SqlInstancesResetReplicaSizeRequest> ) -> Result<Response<Operation>, Status>

Reset Replica Size to primary instance disk size.

pub async fn get_latest_recovery_time( &mut self, request: impl IntoRequest<SqlInstancesGetLatestRecoveryTimeRequest> ) -> Result<Response<SqlInstancesGetLatestRecoveryTimeResponse>, Status>

Get Latest Recovery Time for a given instance.