Module google_api_proto::google::cloud::support::common

Modules

case

Nested message and enum types in Case.

customer_issue

Nested message and enum types in CustomerIssue.

issue_taxonomy

Nested message and enum types in IssueTaxonomy.

support_account

Nested message and enum types in SupportAccount.

support_role

Nested message and enum types in SupportRole.

Structs

Case

A support case created by the user.

Comment

The comment text associated with a Case.

CustomerIssue

Reference to a Google internal ticket used for investigating a support case. Not every support case will have an internal ticket associated with it. A support case can have multiple tickets linked to it.

IssueTaxonomy

Represents the product component taxonomy that is to be used while creating or updating a Case. A client should obtain the list of issue categories, component/subcomponent from this object and specify it in Case.category, Case.component and Case.subcomponent fields respectively.

SupportAccount

A Google Cloud Platform account that identifies support eligibility for a Cloud resource. Currently the Cloud resource can only be an Organization but this might change in future.

SupportRole

A message that contains mapping of a user and their role under a support account.