Module google_api_proto::google::cloud::telcoautomation::v1

Modules

blueprint

Nested message and enum types in Blueprint.

deployment

Nested message and enum types in Deployment.

edge_slm

Nested message and enum types in EdgeSlm.

hydrated_deployment

Nested message and enum types in HydratedDeployment.

management_config

Nested message and enum types in ManagementConfig.

master_authorized_networks_config

Nested message and enum types in MasterAuthorizedNetworksConfig.

orchestration_cluster

Nested message and enum types in OrchestrationCluster.

telco_automation_client

Generated client implementations.

Structs

ApplyDeploymentRequest

Request object for ApplyDeployment. The resources in given deployment gets applied to Orchestration Cluster. A new revision is created when a deployment is applied.

ApplyHydratedDeploymentRequest

Request for applying a hydrated deployment.

ApproveBlueprintRequest

Request object for ApproveBlueprint.

Blueprint

A Blueprint contains a collection of kubernetes resources in the form of YAML files. The file contents of a blueprint are collectively known as package. A blueprint can be a) imported from TNA’s public catalog b) modified as per a user’s need c) proposed and approved. On approval, a revision of blueprint is created which can be used to create a deployment on Orchestration or Workload Cluster.

ComputeDeploymentStatusRequest

Request object for ComputeDeploymentStatus.

ComputeDeploymentStatusResponse

Response object for ComputeDeploymentStatus.

CreateBlueprintRequest

Request object for CreateBlueprint.

CreateDeploymentRequest

Request object for CreateDeployment.

CreateEdgeSlmRequest

Message for creating a EdgeSlm.

CreateOrchestrationClusterRequest

Message for creating a OrchestrationCluster.

DeleteBlueprintRequest

Request object for DeleteBlueprint.

DeleteEdgeSlmRequest

Message for deleting a EdgeSlm.

DeleteOrchestrationClusterRequest

Message for deleting a OrchestrationCluster.

Deployment

Deployment contains a collection of YAML files (This collection is also known as package) that can to applied on an orchestration cluster (GKE cluster with TNA addons) or a workload cluster.

DiscardBlueprintChangesRequest

Request object for DiscardBlueprintChanges.

DiscardBlueprintChangesResponse

Response object for DiscardBlueprintChanges.

DiscardDeploymentChangesRequest

Request object for DiscardDeploymentChanges.

DiscardDeploymentChangesResponse

Response object for DiscardDeploymentChanges.

EdgeSlm

EdgeSlm represents an SLM instance which manages the lifecycle of edge components installed on Workload clusters managed by an Orchestration Cluster.

File

File represents a yaml file present in a blueprint’s package.

FullManagementConfig

Configuration of the full (Autopilot) cluster management

GetBlueprintRequest

Request object for GetBlueprint.

GetDeploymentRequest

Request object for GetDeployment.

GetEdgeSlmRequest

Message for getting a EdgeSlm.

GetHydratedDeploymentRequest

Request object for GetHydratedDeployment.

GetOrchestrationClusterRequest

Message for getting a OrchestrationCluster.

GetPublicBlueprintRequest

Request object for GetPublicBlueprint.

HydratedDeployment

A collection of kubernetes yaml files which are deployed on a Workload Cluster. Hydrated Deployments are created by TNA intent based automation.

HydrationStatus

Hydration status.

ListBlueprintRevisionsRequest

Request object for ListBlueprintRevisions.

ListBlueprintRevisionsResponse

Response object for ListBlueprintRevisions.

ListBlueprintsRequest

Request object for ListBlueprints.

ListBlueprintsResponse

Response object for ListBlueprints.

ListDeploymentRevisionsRequest

Request for listing all revisions of a deployment.

ListDeploymentRevisionsResponse

List of deployment revisions for a given deployment.

ListDeploymentsRequest

Request object for ListDeployments.

ListDeploymentsResponse

Response object for ListDeployments.

ListEdgeSlmsRequest

Message for requesting list of EdgeSlms

ListEdgeSlmsResponse

Message for response to listing EdgeSlms.

ListHydratedDeploymentsRequest

Request object for ListHydratedDeployments.

ListHydratedDeploymentsResponse

Response object for ListHydratedDeployments.

ListOrchestrationClustersRequest

Message for requesting list of OrchestrationClusters.

ListOrchestrationClustersResponse

Message for response to listing OrchestrationClusters.

ListPublicBlueprintsRequest

Request object for ListPublicBlueprints.

ListPublicBlueprintsResponse

Response object for ListPublicBlueprints.

ManagementConfig

Configuration of the cluster management

MasterAuthorizedNetworksConfig

Configuration of the Master Authorized Network that support multiple CIDRs

NfDeploySiteStatus

Per-Site Status.

NfDeployStatus

Deployment status of NFDeploy.

OperationMetadata

Represents the metadata of the long-running operation.

OrchestrationCluster

Orchestration cluster represents a GKE cluster with config controller and TNA specific components installed on it.

ProposeBlueprintRequest

Request object for ProposeBlueprint.

PublicBlueprint

A Blueprint contains a collection of kubernetes resources in the form of YAML files. The file contents of a blueprint are collectively known as package. Public blueprint is a TNA provided blueprint that in present in TNA’s public catalog. A user can copy the public blueprint to their private catalog for further modifications.

RejectBlueprintRequest

Request object for RejectBlueprint.

RemoveDeploymentRequest

Request object for RemoveDeployment.

ResourceStatus

Status of a deployment resource.

RollbackDeploymentRequest

Request object for RollbackDeployment.

SearchBlueprintRevisionsRequest

Request object for SearchBlueprintRevisions.

SearchBlueprintRevisionsResponse

Response object for SearchBlueprintRevisions.

SearchDeploymentRevisionsRequest

Request object for SearchDeploymentRevisions.

SearchDeploymentRevisionsResponse

Response object for SearchDeploymentRevisions.

SiteVersion

SiteVersion Hydration is targeting.

StandardManagementConfig

Configuration of the standard (GKE) cluster management.

UpdateBlueprintRequest

Request object for UpdateBlueprint.

UpdateDeploymentRequest

Request object for UpdateDeployment.

UpdateHydratedDeploymentRequest

Request object for UpdateHydratedDeployment.

WorkloadStatus

Workload status.

Enums

BlueprintView

BlueprintView defines the type of view of the blueprint.

DeploymentLevel

DeploymentLevel of a blueprint signifies where the blueprint will be applied.

DeploymentView

DeploymentView defines the type of view of the deployment.

ResourceType

Represent type of CR.

Status

Status of an entity (resource, deployment).