Module google_api_proto::google::cloud::timeseriesinsights::v1

Modules

data_set

Nested message and enum types in DataSet.

event_dimension

Nested message and enum types in EventDimension.

forecast_params

Nested message and enum types in ForecastParams.

pinned_dimension

Nested message and enum types in PinnedDimension.

timeseries_insights_controller_client

Generated client implementations.

timeseries_params

Nested message and enum types in TimeseriesParams.

Structs

AppendEventsRequest

Appends events to an existing DataSet.

AppendEventsResponse

Response for an AppendEvents RPC.

BigqueryMapping

Mapping of BigQuery columns to timestamp, group_id and dimensions.

CreateDataSetRequest

Create a DataSet request.

DataSet

A collection of data sources sent for processing.

DataSource

A data source consists of multiple [Event][google.cloud.timeseriesinsights.v1.Event] objects stored on Cloud Storage. Each Event should be in JSON format, with one Event per line, also known as JSON Lines format.

DeleteDataSetRequest

Unload DataSet request from the serving system.

EvaluateSliceRequest

Request for evaluateSlice.

EvaluateTimeseriesRequest

Request for evaluateTimeseries.

EvaluatedSlice

Forecast result for a given slice.

Event

Represents an entry in a data source.

EventDimension

Represents an event dimension.

ForecastParams

Parameters that control the sensitivity and other options for the time series forecast.

ListDataSetsRequest

List the DataSets created by the current project.

ListDataSetsResponse

Created DataSets list response.

PinnedDimension

A categorical dimension fixed to a certain value.

QueryDataSetRequest

Request for performing a query against a loaded DataSet.

QueryDataSetResponse

Response for a query executed by the system.

SlicingParams

Parameters that control how we slice the dataset and, optionally, filter slices that have some specific values on some dimensions (pinned dimensions).

Timeseries

A time series.

TimeseriesParams

Parameters that control how we construct the time series for each slice.

TimeseriesPoint

A point in a time series.