Module google_api_proto::google::cloud::tpu::v2alpha1

Modules

attached_disk

Nested message and enum types in AttachedDisk.

node

Nested message and enum types in Node.

symptom

Nested message and enum types in Symptom.

tpu_client

Generated client implementations.

Structs

AcceleratorType

A accelerator type that a Node can be configured with.

AccessConfig

An access config attached to the TPU worker.

AttachedDisk

A node-attached disk resource. Next ID: 8;

CreateNodeRequest

Request for [CreateNode][google.cloud.tpu.v2alpha1.Tpu.CreateNode].

DeleteNodeRequest

Request for [DeleteNode][google.cloud.tpu.v2alpha1.Tpu.DeleteNode].

GenerateServiceIdentityRequest

Request for [GenerateServiceIdentity][google.cloud.tpu.v2alpha1.Tpu.GenerateServiceIdentity].

GenerateServiceIdentityResponse

Response for [GenerateServiceIdentity][google.cloud.tpu.v2alpha1.Tpu.GenerateServiceIdentity].

GetAcceleratorTypeRequest

Request for [GetAcceleratorType][google.cloud.tpu.v2alpha1.Tpu.GetAcceleratorType].

GetGuestAttributesRequest

Request for [GetGuestAttributes][google.cloud.tpu.v2alpha1.Tpu.GetGuestAttributes].

GetGuestAttributesResponse

Response for [GetGuestAttributes][google.cloud.tpu.v2alpha1.Tpu.GetGuestAttributes].

GetNodeRequest

Request for [GetNode][google.cloud.tpu.v2alpha1.Tpu.GetNode].

GetRuntimeVersionRequest

Request for [GetRuntimeVersion][google.cloud.tpu.v2alpha1.Tpu.GetRuntimeVersion].

GuestAttributes

A guest attributes.

GuestAttributesEntry

A guest attributes namespace/key/value entry.

GuestAttributesValue

Array of guest attribute namespace/key/value tuples.

ListAcceleratorTypesRequest

Request for [ListAcceleratorTypes][google.cloud.tpu.v2alpha1.Tpu.ListAcceleratorTypes].

ListAcceleratorTypesResponse

Response for [ListAcceleratorTypes][google.cloud.tpu.v2alpha1.Tpu.ListAcceleratorTypes].

ListNodesRequest

Request for [ListNodes][google.cloud.tpu.v2alpha1.Tpu.ListNodes].

ListNodesResponse

Response for [ListNodes][google.cloud.tpu.v2alpha1.Tpu.ListNodes].

ListRuntimeVersionsRequest

Request for [ListRuntimeVersions][google.cloud.tpu.v2alpha1.Tpu.ListRuntimeVersions].

ListRuntimeVersionsResponse

Response for [ListRuntimeVersions][google.cloud.tpu.v2alpha1.Tpu.ListRuntimeVersions].

NetworkConfig

Network related configurations.

NetworkEndpoint

A network endpoint over which a TPU worker can be reached.

Node

A TPU instance.

OperationMetadata

Metadata describing an [Operation][google.longrunning.Operation]

RuntimeVersion

A runtime version that a Node can be configured with.

SchedulingConfig

Sets the scheduling options for this node.

ServiceAccount

A service account.

ServiceIdentity

The per-product per-project service identity for Cloud TPU service.

StartNodeRequest

Request for [StartNode][google.cloud.tpu.v2alpha1.Tpu.StartNode].

StopNodeRequest

Request for [StopNode][google.cloud.tpu.v2alpha1.Tpu.StopNode].

Symptom

A Symptom instance.

UpdateNodeRequest

Request for [UpdateNode][google.cloud.tpu.v2alpha1.Tpu.UpdateNode].