Module google_api_proto::google::cloud::translation::v3

Modules

batch_document_input_config

Nested message and enum types in BatchDocumentInputConfig.

batch_document_output_config

Nested message and enum types in BatchDocumentOutputConfig.

batch_translate_document_metadata

Nested message and enum types in BatchTranslateDocumentMetadata.

batch_translate_metadata

Nested message and enum types in BatchTranslateMetadata.

create_glossary_metadata

Nested message and enum types in CreateGlossaryMetadata.

delete_glossary_metadata

Nested message and enum types in DeleteGlossaryMetadata.

detect_language_request

Nested message and enum types in DetectLanguageRequest.

document_input_config

Nested message and enum types in DocumentInputConfig.

document_output_config

Nested message and enum types in DocumentOutputConfig.

glossary

Nested message and enum types in Glossary.

glossary_input_config

Nested message and enum types in GlossaryInputConfig.

import_adaptive_mt_file_request

Nested message and enum types in ImportAdaptiveMtFileRequest.

input_config

Nested message and enum types in InputConfig.

output_config

Nested message and enum types in OutputConfig.

translation_service_client

Generated client implementations.

Structs

AdaptiveMtDataset

An Adaptive MT Dataset.

AdaptiveMtFile

An AdaptiveMtFile.

AdaptiveMtSentence

An AdaptiveMt sentence entry.

AdaptiveMtTranslateRequest

The request for sending an AdaptiveMt translation query.

AdaptiveMtTranslateResponse

An AdaptiveMtTranslate response.

AdaptiveMtTranslation

An AdaptiveMt translation.

BatchDocumentInputConfig

Input configuration for BatchTranslateDocument request.

BatchDocumentOutputConfig

Output configuration for BatchTranslateDocument request.

BatchTranslateDocumentMetadata

State metadata for the batch translation operation.

BatchTranslateDocumentRequest

The BatchTranslateDocument request.

BatchTranslateDocumentResponse

Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by BatchTranslateDocument if at least one document is translated successfully.

BatchTranslateMetadata

State metadata for the batch translation operation.

BatchTranslateResponse

Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by BatchTranslateText if at least one sentence is translated successfully.

BatchTranslateTextRequest

The batch translation request.

CreateAdaptiveMtDatasetRequest

Request message for creating an AdaptiveMtDataset.

CreateGlossaryMetadata

Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by CreateGlossary.

CreateGlossaryRequest

Request message for CreateGlossary.

DeleteAdaptiveMtDatasetRequest

Request message for deleting an AdaptiveMtDataset.

DeleteAdaptiveMtFileRequest

The request for deleting an AdaptiveMt file.

DeleteGlossaryMetadata

Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by DeleteGlossary.

DeleteGlossaryRequest

Request message for DeleteGlossary.

DeleteGlossaryResponse

Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by DeleteGlossary.

DetectLanguageRequest

The request message for language detection.

DetectLanguageResponse

The response message for language detection.

DetectedLanguage

The response message for language detection.

DocumentInputConfig

A document translation request input config.

DocumentOutputConfig

A document translation request output config.

DocumentTranslation

A translated document message.

FileInputSource

An inlined file.

GcsDestination

The Google Cloud Storage location for the output content.

GcsInputSource

The Google Cloud Storage location for the input content.

GcsOutputDestination

The Google Cloud Storage location for the output content.

GcsSource

The Google Cloud Storage location for the input content.

GetAdaptiveMtDatasetRequest

Request message for getting an Adaptive MT dataset.

GetAdaptiveMtFileRequest

The request for getting an AdaptiveMtFile.

GetGlossaryRequest

Request message for GetGlossary.

GetSupportedLanguagesRequest

The request message for discovering supported languages.

Glossary

Represents a glossary built from user-provided data.

GlossaryInputConfig

Input configuration for glossaries.

ImportAdaptiveMtFileRequest

The request for importing an AdaptiveMt file along with its sentences.

ImportAdaptiveMtFileResponse

The response for importing an AdaptiveMtFile

InputConfig

Input configuration for BatchTranslateText request.

ListAdaptiveMtDatasetsRequest

Request message for listing all Adaptive MT datasets that the requestor has access to.

ListAdaptiveMtDatasetsResponse

A list of AdaptiveMtDatasets.

ListAdaptiveMtFilesRequest

The request to list all AdaptiveMt files under a given dataset.

ListAdaptiveMtFilesResponse

The response for listing all AdaptiveMt files under a given dataset.

ListAdaptiveMtSentencesRequest

The request for listing Adaptive MT sentences from a Dataset/File.

ListAdaptiveMtSentencesResponse

List AdaptiveMt sentences response.

ListGlossariesRequest

Request message for ListGlossaries.

ListGlossariesResponse

Response message for ListGlossaries.

OutputConfig

Output configuration for BatchTranslateText request.

SupportedLanguage

A single supported language response corresponds to information related to one supported language.

SupportedLanguages

The response message for discovering supported languages.

TranslateDocumentRequest

A document translation request.

TranslateDocumentResponse

A translated document response message.

TranslateTextGlossaryConfig

Configures which glossary is used for a specific target language and defines options for applying that glossary.

TranslateTextRequest

The request message for synchronous translation.

TranslateTextResponse

Translation

A single translation response.