Struct google_api_proto::google::cloud::translation::v3::translation_service_client::TranslationServiceClient

pub struct TranslationServiceClient<T> { /* private fields */ }

Provides natural language translation operations.

Implementations

impl<T> TranslationServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> TranslationServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn translate_text( &mut self, request: impl IntoRequest<TranslateTextRequest> ) -> Result<Response<TranslateTextResponse>, Status>

Translates input text and returns translated text.

pub async fn detect_language( &mut self, request: impl IntoRequest<DetectLanguageRequest> ) -> Result<Response<DetectLanguageResponse>, Status>

Detects the language of text within a request.

pub async fn get_supported_languages( &mut self, request: impl IntoRequest<GetSupportedLanguagesRequest> ) -> Result<Response<SupportedLanguages>, Status>

Returns a list of supported languages for translation.

pub async fn translate_document( &mut self, request: impl IntoRequest<TranslateDocumentRequest> ) -> Result<Response<TranslateDocumentResponse>, Status>

Translates documents in synchronous mode.

pub async fn batch_translate_text( &mut self, request: impl IntoRequest<BatchTranslateTextRequest> ) -> Result<Response<Operation>, Status>

Translates a large volume of text in asynchronous batch mode. This function provides real-time output as the inputs are being processed. If caller cancels a request, the partial results (for an input file, it’s all or nothing) may still be available on the specified output location.

This call returns immediately and you can use google.longrunning.Operation.name to poll the status of the call.

pub async fn batch_translate_document( &mut self, request: impl IntoRequest<BatchTranslateDocumentRequest> ) -> Result<Response<Operation>, Status>

Translates a large volume of document in asynchronous batch mode. This function provides real-time output as the inputs are being processed. If caller cancels a request, the partial results (for an input file, it’s all or nothing) may still be available on the specified output location.

This call returns immediately and you can use google.longrunning.Operation.name to poll the status of the call.

pub async fn create_glossary( &mut self, request: impl IntoRequest<CreateGlossaryRequest> ) -> Result<Response<Operation>, Status>

Creates a glossary and returns the long-running operation. Returns NOT_FOUND, if the project doesn’t exist.

pub async fn list_glossaries( &mut self, request: impl IntoRequest<ListGlossariesRequest> ) -> Result<Response<ListGlossariesResponse>, Status>

Lists glossaries in a project. Returns NOT_FOUND, if the project doesn’t exist.

pub async fn get_glossary( &mut self, request: impl IntoRequest<GetGlossaryRequest> ) -> Result<Response<Glossary>, Status>

Gets a glossary. Returns NOT_FOUND, if the glossary doesn’t exist.

pub async fn delete_glossary( &mut self, request: impl IntoRequest<DeleteGlossaryRequest> ) -> Result<Response<Operation>, Status>

Deletes a glossary, or cancels glossary construction if the glossary isn’t created yet. Returns NOT_FOUND, if the glossary doesn’t exist.

pub async fn create_adaptive_mt_dataset( &mut self, request: impl IntoRequest<CreateAdaptiveMtDatasetRequest> ) -> Result<Response<AdaptiveMtDataset>, Status>

Creates an Adaptive MT dataset.

pub async fn delete_adaptive_mt_dataset( &mut self, request: impl IntoRequest<DeleteAdaptiveMtDatasetRequest> ) -> Result<Response<()>, Status>

Deletes an Adaptive MT dataset, including all its entries and associated metadata.

pub async fn get_adaptive_mt_dataset( &mut self, request: impl IntoRequest<GetAdaptiveMtDatasetRequest> ) -> Result<Response<AdaptiveMtDataset>, Status>

Gets the Adaptive MT dataset.

pub async fn list_adaptive_mt_datasets( &mut self, request: impl IntoRequest<ListAdaptiveMtDatasetsRequest> ) -> Result<Response<ListAdaptiveMtDatasetsResponse>, Status>

Lists all Adaptive MT datasets for which the caller has read permission.

pub async fn adaptive_mt_translate( &mut self, request: impl IntoRequest<AdaptiveMtTranslateRequest> ) -> Result<Response<AdaptiveMtTranslateResponse>, Status>

Translate text using Adaptive MT.

pub async fn get_adaptive_mt_file( &mut self, request: impl IntoRequest<GetAdaptiveMtFileRequest> ) -> Result<Response<AdaptiveMtFile>, Status>

Gets and AdaptiveMtFile

pub async fn delete_adaptive_mt_file( &mut self, request: impl IntoRequest<DeleteAdaptiveMtFileRequest> ) -> Result<Response<()>, Status>

Deletes an AdaptiveMtFile along with its sentences.

pub async fn import_adaptive_mt_file( &mut self, request: impl IntoRequest<ImportAdaptiveMtFileRequest> ) -> Result<Response<ImportAdaptiveMtFileResponse>, Status>

Imports an AdaptiveMtFile and adds all of its sentences into the AdaptiveMtDataset.

pub async fn list_adaptive_mt_files( &mut self, request: impl IntoRequest<ListAdaptiveMtFilesRequest> ) -> Result<Response<ListAdaptiveMtFilesResponse>, Status>

Lists all AdaptiveMtFiles associated to an AdaptiveMtDataset.

pub async fn list_adaptive_mt_sentences( &mut self, request: impl IntoRequest<ListAdaptiveMtSentencesRequest> ) -> Result<Response<ListAdaptiveMtSentencesResponse>, Status>

Lists all AdaptiveMtSentences under a given file/dataset.