Module google_api_proto::google::cloud::video::livestream::v1

Modules

asset

Nested message and enum types in Asset.

audio_stream

Nested message and enum types in AudioStream.

channel

Nested message and enum types in Channel.

elementary_stream

Nested message and enum types in ElementaryStream.

encryption

Nested message and enum types in Encryption.

event

Nested message and enum types in Event.

input

Nested message and enum types in Input.

input_attachment

Nested message and enum types in InputAttachment.

input_config

Nested message and enum types in InputConfig.

livestream_service_client

Generated client implementations.

log_config

Nested message and enum types in LogConfig.

manifest

Nested message and enum types in Manifest.

pool

Nested message and enum types in Pool.

preprocessing_config

Nested message and enum types in PreprocessingConfig.

timecode_config

Nested message and enum types in TimecodeConfig.

video_stream

Nested message and enum types in VideoStream.

Structs

Asset

An asset represents a video or an image.

AudioFormat

Properties of the audio format.

AudioStream

Audio stream resource.

AudioStreamProperty

Properties of the audio stream.

Channel

Channel resource represents the processor that does a user-defined “streaming” operation, which includes getting an input stream through an input, transcoding it to multiple renditions, and publishing output live streams in certain formats (for example, HLS or DASH) to the specified location.

ChannelOperationResponse

Response message for Start/Stop Channel long-running operations.

CreateAssetRequest

Request message for “LivestreamService.CreateAsset”.

CreateChannelRequest

Request message for “LivestreamService.CreateChannel”.

CreateEventRequest

Request message for “LivestreamService.CreateEvent”.

CreateInputRequest

Request message for “LivestreamService.CreateInput”.

DeleteAssetRequest

Request message for “LivestreamService.DeleteAsset”.

DeleteChannelRequest

Request message for “LivestreamService.DeleteChannel”.

DeleteEventRequest

Request message for “LivestreamService.DeleteEvent”.

DeleteInputRequest

Request message for “LivestreamService.DeleteInput”.

ElementaryStream

Encoding of an input element such as an audio, video, or text track. Elementary streams must be packaged before mapping and sharing between different output formats.

Encryption

Encryption settings.

Event

Event is a sub-resource of a channel, which can be scheduled by the user to execute operations on a channel resource without having to stop the channel.

GetAssetRequest

Request message for “LivestreamService.GetAsset”.

GetChannelRequest

Request message for “LivestreamService.GetChannel”.

GetEventRequest

Request message for “LivestreamService.GetEvent”.

GetInputRequest

Request message for “LivestreamService.GetInput”.

GetPoolRequest

Request message for “LivestreamService.GetPool”.

Input

Input resource represents the endpoint from which the channel ingests the input stream.

InputAttachment

A group of information for attaching an input resource to this channel.

InputConfig

Configuration for the input sources of a channel.

InputStreamProperty

Properties of the input stream.

ListAssetsRequest

Request message for “LivestreamService.ListAssets”.

ListAssetsResponse

Response message for “LivestreamService.ListAssets”.

ListChannelsRequest

Request message for “LivestreamService.ListChannels”.

ListChannelsResponse

Response message for “LivestreamService.ListChannels”.

ListEventsRequest

Request message for “LivestreamService.ListEvents”.

ListEventsResponse

Response message for “LivestreamService.ListEvents”.

ListInputsRequest

Request message for “LivestreamService.ListInputs”.

ListInputsResponse

Response message for “LivestreamService.ListInputs”.

LogConfig

Configuration of platform logs. See Using and managing platform logs for more information about how to view platform logs through Cloud Logging.

Manifest

Manifest configuration.

MuxStream

Multiplexing settings for output stream.

OperationMetadata

Represents the metadata of the long-running operation.

Pool

Pool resource defines the configuration of Live Stream pools for a specific location. Currently we support only one pool resource per project per location. After the creation of the first input, a default pool is created automatically at “projects/{project}/locations/{location}/pools/default”.

PreprocessingConfig

Preprocessing configurations.

SegmentSettings

Segment settings for fmp4 and ts.

SpriteSheet

Sprite sheet configuration.

StartChannelRequest

Request message for “LivestreamService.StartChannel”.

StopChannelRequest

Request message for “LivestreamService.StopChannel”.

TextStream

Encoding of a text stream. For example, closed captions or subtitles.

TimecodeConfig

Timecode configuration.

UpdateChannelRequest

Request message for “LivestreamService.UpdateChannel”.

UpdateInputRequest

Request message for “LivestreamService.UpdateInput”.

UpdatePoolRequest

Request message for “LivestreamService.UpdatePool”.

VideoFormat

Properties of the video format.

VideoStream

Video stream resource.

VideoStreamProperty

Properties of the video stream.