Module google_api_proto::google::cloud::video::stitcher::v1

Modules

cdn_key

Nested message and enum types in CdnKey.

companion

Nested message and enum types in Companion.

companion_ads

Nested message and enum types in CompanionAds.

event

Nested message and enum types in Event.

live_session

Nested message and enum types in LiveSession.

manifest_options

Nested message and enum types in ManifestOptions.

video_stitcher_service_client

Generated client implementations.

Structs

AdRequest

Details of an ad request to an ad server.

AdStitchDetail

Metadata for a stitched ad.

AdTag

Metadata of an ad tag.

AkamaiCdnKey

Configuration for an Akamai CDN key.

CdnKey

Configuration for a CDN key. Used by the Video Stitcher to sign URIs for fetching video manifests and signing media segments for playback.

Companion

Metadata for a companion.

CompanionAds

Metadata for companion ads.

CreateCdnKeyRequest

Request message for VideoStitcherService.createCdnKey.

CreateLiveSessionRequest

Request message for VideoStitcherService.createLiveSession.

CreateSlateRequest

Request message for VideoStitcherService.createSlate.

CreateVodSessionRequest

Request message for VideoStitcherService.createVodSession

DeleteCdnKeyRequest

Request message for VideoStitcherService.deleteCdnKey.

DeleteSlateRequest

Request message for VideoStitcherService.deleteSlate.

Event

Describes an event and a trigger URI.

GetCdnKeyRequest

Request message for VideoStitcherService.getCdnKey.

GetLiveAdTagDetailRequest

Request message for VideoStitcherService.getLiveAdTagDetail

GetLiveSessionRequest

Request message for VideoStitcherService.getSession.

GetSlateRequest

Request message for VideoStitcherService.getSlate.

GetVodAdTagDetailRequest

Request message for VideoStitcherService.getVodAdTagDetail

GetVodSessionRequest

Request message for VideoStitcherService.getVodSession

GetVodStitchDetailRequest

Request message for VideoStitcherService.getVodStitchDetail.

GoogleCdnKey

Configuration for a Google Cloud CDN key.

HtmlAdResource

Metadata for an HTML ad resource.

IframeAdResource

Metadata for an IFrame ad resource.

Interstitials

Describes what was stitched into a VOD session’s manifest.

ListCdnKeysRequest

Request message for VideoStitcherService.listCdnKeys.

ListCdnKeysResponse

Response message for VideoStitcher.ListCdnKeys.

ListLiveAdTagDetailsRequest

Request message for VideoStitcherService.listLiveAdTagDetails.

ListLiveAdTagDetailsResponse

Response message for VideoStitcherService.listLiveAdTagDetails.

ListSlatesRequest

Request message for VideoStitcherService.listSlates.

ListSlatesResponse

Response message for VideoStitcherService.listSlates.

ListVodAdTagDetailsRequest

Request message for VideoStitcherService.listVodAdTagDetails.

ListVodAdTagDetailsResponse

Response message for VideoStitcherService.listVodAdTagDetails.

ListVodStitchDetailsRequest

Request message for VideoStitcherService.listVodStitchDetails.

ListVodStitchDetailsResponse

Response message for VideoStitcherService.listVodStitchDetails.

LiveAdTagDetail

Container for a live session’s ad tag detail.

LiveSession

Metadata for a live session.

ManifestOptions

Options for manifest generation.

ProgressEvent

Indicates a time in which a list of events should be triggered during media playback.

RenditionFilter

Filters for a video or muxed redition.

RequestMetadata

Metadata for an ad request.

ResponseMetadata

Metadata for the response of an ad request.

Slate

Slate object

StaticAdResource

Metadata for a static ad resource.

UpdateCdnKeyRequest

Request message for VideoStitcherService.updateCdnKey.

UpdateSlateRequest

Request message for VideoStitcherService.updateSlate.

VodAdTagDetail

Information related to the details for one ad tag.

VodSession

Metadata for a VOD session.

VodSessionAd

Metadata for an inserted ad in a VOD session.

VodSessionAdBreak

Metadata for an inserted ad break.

VodSessionContent

Metadata for the entire stitched content in a VOD session.

VodStitchDetail

Detailed information related to the interstitial of a VOD session.