Struct google_api_proto::google::cloud::video::stitcher::v1::video_stitcher_service_client::VideoStitcherServiceClient

pub struct VideoStitcherServiceClient<T> { /* fields omitted */ }

Video-On-Demand content stitching API allows you to insert ads into (VoD) video on demand files. You will be able to render custom scrubber bars with highlighted ads, enforce ad policies, allow seamless playback and tracking on native players and monetize content with any standard VMAP compliant ad server.

Implementations

impl<T> VideoStitcherServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> VideoStitcherServiceClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn create_cdn_key(
&mut self,
request: impl IntoRequest<CreateCdnKeyRequest>
) -> Result<Response<CdnKey>, Status>

Creates a new CDN key.

pub async fn list_cdn_keys(
&mut self,
request: impl IntoRequest<ListCdnKeysRequest>
) -> Result<Response<ListCdnKeysResponse>, Status>

Lists all CDN keys in the specified project and location.

pub async fn get_cdn_key(
&mut self,
request: impl IntoRequest<GetCdnKeyRequest>
) -> Result<Response<CdnKey>, Status>

Returns the specified CDN key.

pub async fn delete_cdn_key(
&mut self,
request: impl IntoRequest<DeleteCdnKeyRequest>
) -> Result<Response<()>, Status>

Deletes the specified CDN key.

pub async fn update_cdn_key(
&mut self,
request: impl IntoRequest<UpdateCdnKeyRequest>
) -> Result<Response<CdnKey>, Status>

Updates the specified CDN key. Only update fields specified in the call method body.

pub async fn create_vod_session(
&mut self,
request: impl IntoRequest<CreateVodSessionRequest>
) -> Result<Response<VodSession>, Status>

Creates a client side playback VOD session and returns the full tracking and playback metadata of the session.

pub async fn get_vod_session(
&mut self,
request: impl IntoRequest<GetVodSessionRequest>
) -> Result<Response<VodSession>, Status>

Returns the full tracking, playback metadata, and relevant ad-ops logs for the specified VOD session.

pub async fn list_vod_stitch_details(
&mut self,
request: impl IntoRequest<ListVodStitchDetailsRequest>
) -> Result<Response<ListVodStitchDetailsResponse>, Status>

Returns a list of detailed stitching information of the specified VOD session.

pub async fn get_vod_stitch_detail(
&mut self,
request: impl IntoRequest<GetVodStitchDetailRequest>
) -> Result<Response<VodStitchDetail>, Status>

Returns the specified stitching information for the specified VOD session.

pub async fn list_vod_ad_tag_details(
&mut self,
request: impl IntoRequest<ListVodAdTagDetailsRequest>
) -> Result<Response<ListVodAdTagDetailsResponse>, Status>

Return the list of ad tag details for the specified VOD session.

pub async fn get_vod_ad_tag_detail(
&mut self,
request: impl IntoRequest<GetVodAdTagDetailRequest>
) -> Result<Response<VodAdTagDetail>, Status>

Returns the specified ad tag detail for the specified VOD session.

pub async fn list_live_ad_tag_details(
&mut self,
request: impl IntoRequest<ListLiveAdTagDetailsRequest>
) -> Result<Response<ListLiveAdTagDetailsResponse>, Status>

Return the list of ad tag details for the specified live session.

pub async fn get_live_ad_tag_detail(
&mut self,
request: impl IntoRequest<GetLiveAdTagDetailRequest>
) -> Result<Response<LiveAdTagDetail>, Status>

Returns the specified ad tag detail for the specified live session.

pub async fn create_slate(
&mut self,
request: impl IntoRequest<CreateSlateRequest>
) -> Result<Response<Slate>, Status>

Creates a slate.

pub async fn list_slates(
&mut self,
request: impl IntoRequest<ListSlatesRequest>
) -> Result<Response<ListSlatesResponse>, Status>

Lists all slates in the specified project and location.

pub async fn get_slate(
&mut self,
request: impl IntoRequest<GetSlateRequest>
) -> Result<Response<Slate>, Status>

Returns the specified slate.

pub async fn update_slate(
&mut self,
request: impl IntoRequest<UpdateSlateRequest>
) -> Result<Response<Slate>, Status>

Updates the specified slate.

pub async fn delete_slate(
&mut self,
request: impl IntoRequest<DeleteSlateRequest>
) -> Result<Response<()>, Status>

Deletes the specified slate.

pub async fn create_live_session(
&mut self,
request: impl IntoRequest<CreateLiveSessionRequest>
) -> Result<Response<LiveSession>, Status>

Creates a new live session.

pub async fn get_live_session(
&mut self,
request: impl IntoRequest<GetLiveSessionRequest>
) -> Result<Response<LiveSession>, Status>

Returns the details for the specified live session.