Module google_api_proto::google::cloud::videointelligence::v1

Modules

object_tracking_annotation

Nested message and enum types in ObjectTrackingAnnotation.

video_intelligence_service_client

Generated client implementations.

Structs

AnnotateVideoProgress

Video annotation progress. Included in the metadata field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

AnnotateVideoRequest

Video annotation request.

AnnotateVideoResponse

Video annotation response. Included in the response field of the Operation returned by the GetOperation call of the google::longrunning::Operations service.

DetectedAttribute

A generic detected attribute represented by name in string format.

DetectedLandmark

A generic detected landmark represented by name in string format and a 2D location.

Entity

Detected entity from video analysis.

ExplicitContentAnnotation

Explicit content annotation (based on per-frame visual signals only). If no explicit content has been detected in a frame, no annotations are present for that frame.

ExplicitContentDetectionConfig

Config for EXPLICIT_CONTENT_DETECTION.

ExplicitContentFrame

Video frame level annotation results for explicit content.

FaceAnnotation

Deprecated. No effect.

FaceDetectionAnnotation

Face detection annotation.

FaceDetectionConfig

Config for FACE_DETECTION.

FaceFrame

Deprecated. No effect.

FaceSegment

Video segment level annotation results for face detection.

LabelAnnotation

Label annotation.

LabelDetectionConfig

Config for LABEL_DETECTION.

LabelFrame

Video frame level annotation results for label detection.

LabelSegment

Video segment level annotation results for label detection.

LogoRecognitionAnnotation

Annotation corresponding to one detected, tracked and recognized logo class.

NormalizedBoundingBox

Normalized bounding box. The normalized vertex coordinates are relative to the original image. Range: [0, 1].

NormalizedBoundingPoly

Normalized bounding polygon for text (that might not be aligned with axis). Contains list of the corner points in clockwise order starting from top-left corner. For example, for a rectangular bounding box: When the text is horizontal it might look like: 0––1 | | 3––2

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ObjectTrackingAnnotation

Annotations corresponding to one tracked object.

ObjectTrackingConfig

Config for OBJECT_TRACKING.

ObjectTrackingFrame

Video frame level annotations for object detection and tracking. This field stores per frame location, time offset, and confidence.

PersonDetectionAnnotation

Person detection annotation per video.

PersonDetectionConfig

Config for PERSON_DETECTION.

ShotChangeDetectionConfig

Config for SHOT_CHANGE_DETECTION.

SpeechContext

Provides “hints” to the speech recognizer to favor specific words and phrases in the results.

SpeechRecognitionAlternative

Alternative hypotheses (a.k.a. n-best list).

SpeechTranscription

A speech recognition result corresponding to a portion of the audio.

SpeechTranscriptionConfig

Config for SPEECH_TRANSCRIPTION.

TextAnnotation

Annotations related to one detected OCR text snippet. This will contain the corresponding text, confidence value, and frame level information for each detection.

TextDetectionConfig

Config for TEXT_DETECTION.

TextFrame

Video frame level annotation results for text annotation (OCR). Contains information regarding timestamp and bounding box locations for the frames containing detected OCR text snippets.

TextSegment

Video segment level annotation results for text detection.

TimestampedObject

For tracking related features. An object at time_offset with attributes, and located with normalized_bounding_box.

Track

A track of an object instance.

VideoAnnotationProgress

Annotation progress for a single video.

VideoAnnotationResults

Annotation results for a single video.

VideoContext

Video context and/or feature-specific parameters.

VideoSegment

Video segment.

WordInfo

Word-specific information for recognized words. Word information is only included in the response when certain request parameters are set, such as enable_word_time_offsets.

Enums

Feature

Video annotation feature.

LabelDetectionMode

Label detection mode.

Likelihood

Bucketized representation of likelihood.