Module google_api_proto::google::cloud::visionai::v1

Modules

analyze_asset_metadata

Nested message and enum types in AnalyzeAssetMetadata.

analyzer_definition

Nested message and enum types in AnalyzerDefinition.

annotation_value

Nested message and enum types in AnnotationValue.

app_platform_client

Generated client implementations.

app_platform_cloud_function_request

Nested message and enum types in AppPlatformCloudFunctionRequest.

app_platform_cloud_function_response

Nested message and enum types in AppPlatformCloudFunctionResponse.

application

Nested message and enum types in Application.

application_configs

Nested message and enum types in ApplicationConfigs.

asset_source

Nested message and enum types in AssetSource.

attribute_value

Nested message and enum types in AttributeValue.

batch_run_process_request

Nested message and enum types in BatchRunProcessRequest.

clip_asset_response

Nested message and enum types in ClipAssetResponse.

cluster

Nested message and enum types in Cluster.

collection_item

Nested message and enum types in CollectionItem.

controlled_mode

Nested message and enum types in ControlledMode.

corpus

Nested message and enum types in Corpus.

criteria

Nested message and enum types in Criteria.

custom_processor_source_info

Nested message and enum types in CustomProcessorSourceInfo.

data_schema_details

Nested message and enum types in DataSchemaDetails.

event

Nested message and enum types in Event.

facet_bucket

Nested message and enum types in FacetBucket.

facet_property

Nested message and enum types in FacetProperty.

facet_value

Nested message and enum types in FacetValue.

health_check_service_client

Generated client implementations.

image_query

Nested message and enum types in ImageQuery.

import_assets_request

Nested message and enum types in ImportAssetsRequest.

index

Nested message and enum types in Index.

index_endpoint

Nested message and enum types in IndexEndpoint.

indexing_status

Nested message and enum types in IndexingStatus.

ingest_asset_request

Nested message and enum types in IngestAssetRequest.

instance

Nested message and enum types in Instance.

live_video_analytics_client

Generated client implementations.

node

Nested message and enum types in Node.

object_detection_prediction_result

Nested message and enum types in ObjectDetectionPredictionResult.

occupancy_counting_prediction_result

Nested message and enum types in OccupancyCountingPredictionResult.

operator_definition

Nested message and enum types in OperatorDefinition.

packet_type

Nested message and enum types in PacketType.

partition

Nested message and enum types in Partition.

person_blur_config

Nested message and enum types in PersonBlurConfig.

personal_protective_equipment_detection_output

Nested message and enum types in PersonalProtectiveEquipmentDetectionOutput.

prediction_service_client

Generated client implementations.

processor

Nested message and enum types in Processor.

processor_config

Nested message and enum types in ProcessorConfig.

processor_io_spec

Nested message and enum types in ProcessorIOSpec.

receive_events_control_response

Nested message and enum types in ReceiveEventsControlResponse.

receive_events_request

Nested message and enum types in ReceiveEventsRequest.

receive_events_response

Nested message and enum types in ReceiveEventsResponse.

receive_packets_control_response

Nested message and enum types in ReceivePacketsControlResponse.

receive_packets_request

Nested message and enum types in ReceivePacketsRequest.

receive_packets_response

Nested message and enum types in ReceivePacketsResponse.

remove_application_stream_input_request

Nested message and enum types in RemoveApplicationStreamInputRequest.

run_status

Nested message and enum types in RunStatus.

schema_key_sorting_strategy

Nested message and enum types in SchemaKeySortingStrategy.

search_assets_request

Nested message and enum types in SearchAssetsRequest.

search_capability

Nested message and enum types in SearchCapability.

search_index_endpoint_request

Nested message and enum types in SearchIndexEndpointRequest.

send_packets_request

Nested message and enum types in SendPacketsRequest.

stream_annotation

Nested message and enum types in StreamAnnotation.

stream_with_annotation

Nested message and enum types in StreamWithAnnotation.

streaming_service_client

Generated client implementations.

streams_service_client

Generated client implementations.

tag_parsing_config

Nested message and enum types in TagParsingConfig.

update_application_instances_request

Nested message and enum types in UpdateApplicationInstancesRequest.

video_action_recognition_prediction_result

Nested message and enum types in VideoActionRecognitionPredictionResult.

video_classification_prediction_result

Nested message and enum types in VideoClassificationPredictionResult.

video_object_tracking_prediction_result

Nested message and enum types in VideoObjectTrackingPredictionResult.

warehouse_client

Generated client implementations.

Structs

AcquireLeaseRequest

Request message for acquiring a lease.

AddApplicationStreamInputRequest

Message for adding stream input to an Application.

AddApplicationStreamInputResponse

Message for AddApplicationStreamInput Response.

AddCollectionItemRequest

Request message for AddCollectionItem.

AddCollectionItemResponse

Response message for AddCollectionItem.

AiEnabledDevicesInputConfig

Message describing AI-enabled Devices Input Config.

Analysis

Message describing the Analysis object.

AnalysisDefinition

Defines a full analysis.

AnalyzeAssetMetadata

Metadata for AnalyzeAsset.

AnalyzeAssetRequest

Request message for AnalyzeAsset.

AnalyzeAssetResponse

Response message for AnalyzeAsset.

AnalyzeCorpusMetadata

The metadata message for AnalyzeCorpus LRO.

AnalyzeCorpusRequest

Request message for AnalyzeCorpus.

AnalyzeCorpusResponse

The response message for AnalyzeCorpus LRO.

AnalyzerDefinition

Defines an Analyzer.

Annotation

An annotation is a resource in asset. It represents a key-value mapping of content in asset.

AnnotationCustomizedStruct

Customized struct represnation in annotation.

AnnotationList

List representation in annotation.

AnnotationMatchingResult

Stores the criteria-annotation matching results for each search result item.

AnnotationValue

Value of annotation, including all types available in data schema.

AppPlatformCloudFunctionRequest

For any cloud function based customer processing logic, customer’s cloud function is expected to receive AppPlatformCloudFunctionRequest as request and send back AppPlatformCloudFunctionResponse as response. Message of request from AppPlatform to Cloud Function.

AppPlatformCloudFunctionResponse

Message of the response from customer’s Cloud Function to AppPlatform.

AppPlatformEventBody

Message of content of appPlatform event

AppPlatformMetadata

Message of essential metadata of App Platform. This message is usually attached to a certain processor output annotation for customer to identify the source of the data.

Application

Message describing Application object

ApplicationConfigs

Message storing the graph of the application.

ApplicationInstance

Message for creating a Instance.

ApplicationNodeAnnotation

Message describing annotations specific to application node. This message is a duplication of StreamWithAnnotation.NodeAnnotation.

ApplicationStreamInput

Message about a single stream input config.

Asset

An asset is a resource in corpus. It represents a media object inside corpus, contains metadata and another resource annotation. Different feature could be applied to the asset to generate annotations. User could specified annotation related to the target asset.

AssetSource

The source of the asset.

AttributeValue

Represents an actual value of an operator attribute.

AutoscalingMetricSpec

The metric specification that defines the target resource utilization (CPU utilization, accelerator’s duty cycle, and so on) for calculating the desired replica count.

BatchRunProcessRequest

Request message for running the processes in a batch.

BatchRunProcessResponse

Response message for running the processes in a batch.

BigQueryConfig

Message of configurations for BigQuery processor.

BoolValue

Channel

Message describing the Channel object.

CircleArea

Representation of a circle area.

ClassificationPredictionResult

Prediction output format for Image and Text Classification.

ClipAssetRequest

Request message for ClipAsset API.

ClipAssetResponse

Response message for ClipAsset API.

Cluster

Message describing the Cluster object.

ClusterInfo

Collection

A collection is a resource in a corpus. It serves as a container of references to original resources.

CollectionItem

A CollectionItem is an item in a collection. Each item is a reference to the original resource in a collection.

CommitRequest

The message for explicitly committing the read progress.

ControlledMode

The options for receiver under the controlled mode.

Corpus

Corpus is a set of media contents for management. Within a corpus, media shares the same data schema. Search is also restricted within a single corpus.

CreateAnalysisRequest

Message for creating an Analysis.

CreateAnnotationRequest

Request message for CreateAnnotation.

CreateApplicationInstancesRequest

Message for adding stream input to an Application.

CreateApplicationInstancesResponse

Message for CreateApplicationInstance Response.

CreateApplicationRequest

Message for creating a Application.

CreateAssetRequest

Request message for CreateAssetRequest.

CreateClusterRequest

Message for creating a Cluster.

CreateCollectionMetadata

Metadata message for CreateCollectionRequest

CreateCollectionRequest

Request message for CreateCollection.

CreateCorpusMetadata

Metadata for CreateCorpus API.

CreateCorpusRequest

Request message of CreateCorpus API.

CreateDataSchemaRequest

Request message for CreateDataSchema.

CreateDraftRequest

Message for creating a Draft.

CreateEventRequest

Message for creating a Event.

CreateIndexEndpointMetadata

Metadata message for CreateIndexEndpoint.

CreateIndexEndpointRequest

Request message for CreateIndexEndpoint.

CreateIndexMetadata

Metadata message for CreateIndexRequest

CreateIndexRequest

Message for creating an Index.

CreateOperatorRequest

Message for creating a Operator.

CreateProcessRequest

Message for creating a Process.

CreateProcessorRequest

Message for creating a Processor.

CreateSearchConfigRequest

Request message for CreateSearchConfig.

CreateSearchHypernymRequest

Request message for creating SearchHypernym.

CreateSeriesRequest

Message for creating a Series.

CreateStreamRequest

Message for creating a Stream.

Criteria

Filter criteria applied to current search results.

CustomProcessorSourceInfo

Describes the source info for a custom processor.

DataSchema

Data schema indicates how the user specified annotation is interpreted in the system.

DataSchemaDetails

Data schema details indicates the data type and the data struct corresponding to the key of user specified annotation.

DateTimeRange

Datetime range type.

DateTimeRangeArray

A list of datetime range values.

DedicatedResources

A description of resources that are dedicated to a DeployedModel, and that need a higher degree of manual configuration.

DeleteAnalysisRequest

Message for deleting an Analysis.

DeleteAnnotationRequest

Request message for DeleteAnnotation API.

DeleteApplicationInstancesRequest

Message for removing stream input from an Application.

DeleteApplicationInstancesResponse

Message for DeleteApplicationInstance Response.

DeleteApplicationRequest

Message for deleting an Application.

DeleteAssetMetadata

The metadata for DeleteAsset API that embeds in [metadata][google.longrunning.Operation.metadata] field.

DeleteAssetRequest

Request message for DeleteAsset.

DeleteClusterRequest

Message for deleting a Cluster.

DeleteCollectionMetadata

Metadata message for DeleteCollectionRequest

DeleteCollectionRequest

Request message for DeleteCollectionRequest.

DeleteCorpusRequest

Request message for DeleteCorpus.

DeleteDataSchemaRequest

Request message for DeleteDataSchema.

DeleteDraftRequest

Message for deleting a Draft.

DeleteEventRequest

Message for deleting a Event.

DeleteIndexEndpointMetadata

Metadata message for DeleteIndexEndpoint.

DeleteIndexEndpointRequest

Request message for DeleteIndexEndpoint.

DeleteIndexMetadata

Metadata message for DeleteIndexRequest

DeleteIndexRequest

Request message for DeleteIndex.

DeleteOperatorRequest

Message for deleting a Operator

DeleteProcessRequest

Message for deleting a Process.

DeleteProcessorRequest

Message for deleting a Processor.

DeleteSearchConfigRequest

Request message for DeleteSearchConfig.

DeleteSearchHypernymRequest

Request message for deleting SearchHypernym.

DeleteSeriesRequest

Message for deleting a Series.

DeleteStreamRequest

Message for deleting a Stream.

DeployApplicationRequest

Message for deploying an Application.

DeployApplicationResponse

RPC Request Messages. Message for DeployApplication Response.

DeployIndexMetadata

Metadata message for DeployIndex.

DeployIndexRequest

Request message for DeployIndex.

DeployIndexResponse

DeployIndex response once the operation is done.

DeployedIndex

A deployment of an Index.

DeployedIndexReference

Points to a DeployedIndex.

Draft

Message describing Draft object

EagerMode

The options for receiver under the eager mode.

Event

Message describing the Event object.

EventUpdate

The event update message.

FacetBucket

Holds the facet value, selections state, and metadata.

FacetGroup

A group of facet buckets to be passed back and forth between backend & frontend.

FacetProperty

Central configuration for a facet.

FacetValue

Definition of a single value with generic type.

FloatRange

Float range type.

FloatRangeArray

A list of float range values.

GcsOutputConfig

Message describing GcsOutputConfig.

GcsSource

The Google Cloud Storage location for the input content.

GeneralObjectDetectionConfig

Message of configurations for General Object Detection processor.

GenerateHlsUriRequest

Request message for GenerateHlsUri API.

GenerateHlsUriResponse

Response message for GenerateHlsUri API.

GenerateRetrievalUrlRequest

Request message for GenerateRetrievalUrl API.

GenerateRetrievalUrlResponse

Response message for GenerateRetrievalUrl API.

GenerateStreamHlsTokenRequest

Request message for getting the auth token to access the stream HLS contents.

GenerateStreamHlsTokenResponse

Response message for GenerateStreamHlsToken.

GeoCoordinate

Location Coordinate Representation

GeoLocationArray

A list of locations.

GetAnalysisRequest

Message for getting an Analysis.

GetAnnotationRequest

Request message for GetAnnotation API.

GetApplicationRequest

Message for getting a Application.

GetAssetRequest

Request message for GetAsset.

GetClusterRequest

Message for getting a Cluster.

GetCollectionRequest

Request message for GetCollectionRequest.

GetCorpusRequest

Request message for GetCorpus.

GetDataSchemaRequest

Request message for GetDataSchema.

GetDraftRequest

Message for getting a Draft.

GetEventRequest

Message for getting a Event.

GetIndexEndpointRequest

Request message for GetIndexEndpoint.

GetIndexRequest

Request message for getting an Index.

GetInstanceRequest

Message for getting a Instance.

GetOperatorRequest

Message for getting a Operator.

GetProcessRequest

Message for getting a Process.

GetProcessorRequest

Message for getting a Processor.

GetSearchConfigRequest

Request message for GetSearchConfig.

GetSearchHypernymRequest

Request message for getting SearchHypernym.

GetSeriesRequest

Message for getting a Series.

GetStreamRequest

Message for getting a Stream.

GetStreamThumbnailRequest

Message for getting the thumbnail of a Stream.

GetStreamThumbnailResponse

Message for the response of GetStreamThumbnail. The empty response message indicates the thumbnail image has been uploaded to GCS successfully.

GstreamerBufferDescriptor

The descriptor for a gstreamer buffer payload.

HealthCheckRequest

HealthCheckRequest is the request message for Check.

HealthCheckResponse

HealthCheckResponse is the response message for Check.

ImageObjectDetectionPredictionResult

Prediction output format for Image Object Detection.

ImageQuery

Image query for search endpoint request.

ImageSegmentationPredictionResult

Prediction output format for Image Segmentation.

ImportAssetsMetadata

The metadata message for ImportAssets LRO.

ImportAssetsRequest

The request message for ImportAssets.

ImportAssetsResponse

The response message for ImportAssets LRO.

Index

An Index is a resource in Corpus. It contains an indexed version of the assets and annotations. When deployed to an endpoint, it will allow users to search the Index.

IndexAssetMetadata

Metadata for IndexAsset.

IndexAssetRequest

Request message for IndexAsset.

IndexAssetResponse

Response message for IndexAsset.

IndexEndpoint

Message representing IndexEndpoint resource. Indexes are deployed into it.

IndexedAsset

An IndexedAsset is an asset that the index is built upon.

IndexingStatus

The status of indexing for the asset.

IngestAssetRequest

Request message for IngestAsset API.

IngestAssetResponse

Response message for IngestAsset API.

Instance

Message describing Instance object Next ID: 12

IntRange

Integer range type.

IntRangeArray

A list of integer range values.

Lease

The lease message.

ListAnalysesRequest

Message for requesting list of Analyses

ListAnalysesResponse

Message for response to listing Analyses

ListAnnotationsRequest

Request message for GetAnnotation API.

ListAnnotationsResponse

Request message for ListAnnotations API.

ListApplicationsRequest

Message for requesting list of Applications.

ListApplicationsResponse

Message for response to listing Applications.

ListAssetsRequest

Request message for ListAssets.

ListAssetsResponse

Response message for ListAssets.

ListClustersRequest

Message for requesting list of Clusters.

ListClustersResponse

Message for response to listing Clusters.

ListCollectionsRequest

Request message for ListCollections.

ListCollectionsResponse

Response message for ListCollections.

ListCorporaRequest

Request message for ListCorpora.

ListCorporaResponse

Response message for ListCorpora.

ListDataSchemasRequest

Request message for ListDataSchemas.

ListDataSchemasResponse

Response message for ListDataSchemas.

ListDraftsRequest

Message for requesting list of Drafts.

ListDraftsResponse

Message for response to listing Drafts.

ListEventsRequest

Message for requesting list of Events.

ListEventsResponse

Message for response to listing Events.

ListIndexEndpointsRequest

Request message for ListIndexEndpoints.

ListIndexEndpointsResponse

Response message for ListIndexEndpoints.

ListIndexesRequest

Request message for listing Indexes.

ListIndexesResponse

Response message for ListIndexes.

ListInstancesRequest

Message for requesting list of Instances.

ListInstancesResponse

Message for response to listing Instances.

ListOperatorsRequest

Message for requesting list of Operators.

ListOperatorsResponse

Message for response to listing Operators.

ListPrebuiltProcessorsRequest

Request Message for listing Prebuilt Processors.

ListPrebuiltProcessorsResponse

Response Message for listing Prebuilt Processors.

ListProcessesRequest

Message for requesting list of Processes.

ListProcessesResponse

Message for response to listing Processes.

ListProcessorsRequest

Message for requesting list of Processors.

ListProcessorsResponse

Message for response to listing Processors.

ListPublicOperatorsRequest

Request message of ListPublicOperatorsRequest API.

ListPublicOperatorsResponse

Response message of ListPublicOperators API.

ListSearchConfigsRequest

Request message for ListSearchConfigs.

ListSearchConfigsResponse

Response message for ListSearchConfigs.

ListSearchHypernymsRequest

Request message for listing SearchHypernyms.

ListSearchHypernymsResponse

Response message for listing SearchHypernyms.

ListSeriesRequest

Message for requesting list of Series.

ListSeriesResponse

Message for response to listing Series.

ListStreamsRequest

Message for requesting list of Streams.

ListStreamsResponse

Message for response to listing Streams.

MachineSpec

Specification of a single machine.

MaterializeChannelRequest

Message for materializing a channel.

MediaWarehouseConfig

Message describing MediaWarehouseConfig.

Node

Message describing node object.

NormalizedPolygon

Normalized Polygon.

NormalizedPolyline

Normalized Pplyline, which represents a curve consisting of connected straight-line segments.

NormalizedVertex

A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ObjectDetectionPredictionResult

Prediction output format for Generic Object Detection.

OccupancyCountConfig

Message describing OccupancyCountConfig.

OccupancyCountingPredictionResult

The prediction result proto for occupancy counting.

OperationMetadata

Represents the metadata of the long-running operation.

Operator

Message describing the Operator object.

OperatorDefinition

Defines the interface of an Operator.

OperatorQuery

OperatorQuery represents one query to a Operator.

Packet

The quanta of datum that the series accepts.

PacketHeader

The message that represents packet header.

PacketType

The message that represents the data type of a packet.

Partition

Partition to specify the partition in time and space for sub-asset level annotation.

PersonBlurConfig

Message describing FaceBlurConfig.

PersonVehicleDetectionConfig

Message describing PersonVehicleDetectionConfig.

PersonalProtectiveEquipmentDetectionConfig

Message describing PersonalProtectiveEquipmentDetectionConfig.

PersonalProtectiveEquipmentDetectionOutput

Output format for Personal Protective Equipment Detection Operator.

Process

Message describing the Process object.

Processor

Message describing Processor object. Next ID: 19

ProcessorConfig

Next ID: 35

ProcessorIoSpec

Message describing the input / output specifications of a processor.

ProductRecognizerConfig

Message describing ProductRecognizerConfig.

RawImageDescriptor

The descriptor for a raw image.

ReceiveEventsControlResponse

Control message for a ReceiveEventsResponse.

ReceiveEventsRequest

Request message for ReceiveEvents.

ReceiveEventsResponse

Response message for the ReceiveEvents.

ReceivePacketsControlResponse

Control message for a ReceivePacketsResponse.

ReceivePacketsRequest

Request message for receiving packets.

ReceivePacketsResponse

Response message from ReceivePackets.

ReleaseLeaseRequest

Request message for releasing lease.

ReleaseLeaseResponse

Response message for release lease.

RemoveApplicationStreamInputRequest

Message for removing stream input from an Application.

RemoveApplicationStreamInputResponse

Message for RemoveApplicationStreamInput Response.

RemoveCollectionItemRequest

Request message for RemoveCollectionItem.

RemoveCollectionItemResponse

Request message for RemoveCollectionItem.

RemoveIndexAssetMetadata

Metadata for RemoveIndexAsset.

RemoveIndexAssetRequest

Request message for RemoveIndexAsset.

RemoveIndexAssetResponse

Response message for RemoveIndexAsset.

RenewLeaseRequest

Request message for renewing a lease.

RequestMetadata

RequestMetadata is the metadata message for the request.

ResolveOperatorInfoRequest

Request message for querying operator info.

ResolveOperatorInfoResponse

Response message of ResolveOperatorInfo API.

ResourceAnnotations

Message describing general annotation for resources.

ResourceSpecification

ResourceSpec collects a set of resources that can be used to specify requests and requirements.

RunStatus

Message describing the status of the Process.

SchemaKeySortingStrategy

A strategy to specify how to sort by data schema key.

SearchAssetsRequest

Request message for SearchAssets.

SearchAssetsResponse

Response message for SearchAssets.

SearchCapability

The capability and metadata of search capability.

SearchCapabilitySetting

Setting for search capability to enable.

SearchConfig

SearchConfig stores different properties that will affect search behaviors and search results.

SearchCriteriaProperty

Central configuration for custom search criteria.

SearchHypernym

Search resource: SearchHypernym. For example, { hypernym: “vehicle” hyponyms: [“sedan”, “truck”] } This means in SMART_SEARCH mode, searching for “vehicle” will also return results with “sedan” or “truck” as annotations.

SearchIndexEndpointRequest

Request message for SearchIndexEndpoint.

SearchIndexEndpointResponse

Response message for SearchIndexEndpoint.

SearchResultItem

Search result contains asset name and corresponding time ranges.

SendPacketsRequest

Request message for sending packets.

SendPacketsResponse

Response message for sending packets.

Series

Message describing the Series object.

SeriesMetadata

The message that represents series metadata.

ServerMetadata

The message that represents server metadata.

Stream

Message describing the Stream object. The Stream and the Event resources are many to many; i.e., each Stream resource can associate to many Event resources and each Event resource can associate to many Stream resources.

StreamAnnotation

message about annotations about Vision AI stream resource.

StreamAnnotations

A wrapper of repeated StreamAnnotation.

StreamWithAnnotation

Message describing Vision AI stream with application specific annotations. All the StreamAnnotation object inside this message MUST have unique id.

StringArray

A list of string-type values.

TagParsingConfig

Configuration for tag parsing.

TagRecognizerConfig

Message describing TagRecognizerConfig.

UndeployApplicationRequest

Message for undeploying an Application.

UndeployApplicationResponse

Message for UndeployApplication Response.

UndeployIndexMetadata

Metadata message for UndeployIndex.

UndeployIndexRequest

Request message for UndeployIndexEndpoint.

UndeployIndexResponse

UndeployIndex response once the operation is done.

UniversalInputConfig

Message describing UniversalInputConfig.

UpdateAnalysisRequest

Message for updating an Analysis.

UpdateAnnotationRequest

Request message for UpdateAnnotation API.

UpdateApplicationInstancesRequest

Message for updating an ApplicationInstance.

UpdateApplicationInstancesResponse

Message for UpdateApplicationInstances Response.

UpdateApplicationRequest

Message for updating an Application.

UpdateApplicationStreamInputRequest

Message for updating stream input to an Application.

UpdateApplicationStreamInputResponse

Message for AddApplicationStreamInput Response.

UpdateAssetRequest

Request message for UpdateAsset.

UpdateClusterRequest

Message for updating a Cluster.

UpdateCollectionRequest

Request message for UpdateCollectionRequest.

UpdateCorpusRequest

Request message for UpdateCorpus.

UpdateDataSchemaRequest

Request message for UpdateDataSchema.

UpdateDraftRequest

Message for updating a Draft.

UpdateEventRequest

Message for updating a Event.

UpdateIndexEndpointMetadata

Metadata message for UpdateIndexEndpoint.

UpdateIndexEndpointRequest

Request message for UpdateIndexEndpoint.

UpdateIndexMetadata

Metadata message for UpdateIndexRequest

UpdateIndexRequest

Request message for UpdateIndex.

UpdateOperatorRequest

Message for updating a Operator.

UpdateProcessRequest

Message for updating a Process.

UpdateProcessorRequest

Message for updating a Processor.

UpdateSearchConfigRequest

Request message for UpdateSearchConfig.

UpdateSearchHypernymRequest

Request message for updating SearchHypernym.

UpdateSeriesRequest

Message for updating a Series.

UpdateStreamRequest

Message for updating a Stream.

UploadAssetMetadata

Metadata for UploadAsset.

UploadAssetRequest

Request message for UploadAsset.

UploadAssetResponse

Response message for UploadAsset.

UserSpecifiedAnnotation

Annotation provided by users.

VertexAutoMlVideoConfig

Message describing VertexAutoMLVideoConfig.

VertexAutoMlVisionConfig

Message of configurations of Vertex AutoML Vision Processors.

VertexCustomConfig

Message describing VertexCustomConfig.

VideoActionRecognitionPredictionResult

Prediction output format for Video Action Recognition.

VideoClassificationPredictionResult

Prediction output format for Video Classification.

VideoObjectTrackingPredictionResult

Prediction output format for Video Object Tracking.

VideoStreamInputConfig

Message describing Video Stream Input Config. This message should only be used as a placeholder for builtin:stream-input processor, actual stream binding should be specified using corresponding API.

ViewCollectionItemsRequest

Request message for ViewCollectionItems.

ViewCollectionItemsResponse

Response message for ViewCollectionItems.

ViewIndexedAssetsRequest

Request message for ViewIndexedAssets.

ViewIndexedAssetsResponse

Response message for ViewIndexedAssets.

Enums

AcceleratorType

Represents a hardware accelerator type.

DataType

All supported data types.

FacetBucketType

Different types for a facet bucket.

LeaseType

The lease type.

ModelType

All the supported model types in Vision AI App Platform.

Registry

The enum of the types of the Registry.

RunMode

RunMode represents the mode to launch the Process on.

StreamAnnotationType

Enum describing all possible types of a stream annotation.