Struct google_api_proto::google::cloud::vmmigration::v1::vm_migration_client::VmMigrationClient

pub struct VmMigrationClient<T> { /* fields omitted */ }

VM Migration Service

Implementations

impl<T> VmMigrationClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> VmMigrationClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn list_sources(
&mut self,
request: impl IntoRequest<ListSourcesRequest>
) -> Result<Response<ListSourcesResponse>, Status>

Lists Sources in a given project and location.

pub async fn get_source(
&mut self,
request: impl IntoRequest<GetSourceRequest>
) -> Result<Response<Source>, Status>

Gets details of a single Source.

pub async fn create_source(
&mut self,
request: impl IntoRequest<CreateSourceRequest>
) -> Result<Response<Operation>, Status>

Creates a new Source in a given project and location.

pub async fn update_source(
&mut self,
request: impl IntoRequest<UpdateSourceRequest>
) -> Result<Response<Operation>, Status>

Updates the parameters of a single Source.

pub async fn delete_source(
&mut self,
request: impl IntoRequest<DeleteSourceRequest>
) -> Result<Response<Operation>, Status>

Deletes a single Source.

pub async fn fetch_inventory(
&mut self,
request: impl IntoRequest<FetchInventoryRequest>
) -> Result<Response<FetchInventoryResponse>, Status>

List remote source’s inventory of VMs. The remote source is the onprem vCenter (remote in the sense it’s not in Compute Engine). The inventory describes the list of existing VMs in that source. Note that this operation lists the VMs on the remote source, as opposed to listing the MigratingVms resources in the vmmigration service.

pub async fn list_utilization_reports(
&mut self,
request: impl IntoRequest<ListUtilizationReportsRequest>
) -> Result<Response<ListUtilizationReportsResponse>, Status>

Lists Utilization Reports of the given Source.

pub async fn get_utilization_report(
&mut self,
request: impl IntoRequest<GetUtilizationReportRequest>
) -> Result<Response<UtilizationReport>, Status>

Gets a single Utilization Report.

pub async fn create_utilization_report(
&mut self,
request: impl IntoRequest<CreateUtilizationReportRequest>
) -> Result<Response<Operation>, Status>

Creates a new UtilizationReport.

pub async fn delete_utilization_report(
&mut self,
request: impl IntoRequest<DeleteUtilizationReportRequest>
) -> Result<Response<Operation>, Status>

Deletes a single Utilization Report.

pub async fn list_datacenter_connectors(
&mut self,
request: impl IntoRequest<ListDatacenterConnectorsRequest>
) -> Result<Response<ListDatacenterConnectorsResponse>, Status>

Lists DatacenterConnectors in a given Source.

pub async fn get_datacenter_connector(
&mut self,
request: impl IntoRequest<GetDatacenterConnectorRequest>
) -> Result<Response<DatacenterConnector>, Status>

Gets details of a single DatacenterConnector.

pub async fn create_datacenter_connector(
&mut self,
request: impl IntoRequest<CreateDatacenterConnectorRequest>
) -> Result<Response<Operation>, Status>

Creates a new DatacenterConnector in a given Source.

pub async fn delete_datacenter_connector(
&mut self,
request: impl IntoRequest<DeleteDatacenterConnectorRequest>
) -> Result<Response<Operation>, Status>

Deletes a single DatacenterConnector.

pub async fn create_migrating_vm(
&mut self,
request: impl IntoRequest<CreateMigratingVmRequest>
) -> Result<Response<Operation>, Status>

Creates a new MigratingVm in a given Source.

pub async fn list_migrating_vms(
&mut self,
request: impl IntoRequest<ListMigratingVmsRequest>
) -> Result<Response<ListMigratingVmsResponse>, Status>

Lists MigratingVms in a given Source.

pub async fn get_migrating_vm(
&mut self,
request: impl IntoRequest<GetMigratingVmRequest>
) -> Result<Response<MigratingVm>, Status>

Gets details of a single MigratingVm.

pub async fn update_migrating_vm(
&mut self,
request: impl IntoRequest<UpdateMigratingVmRequest>
) -> Result<Response<Operation>, Status>

Updates the parameters of a single MigratingVm.

pub async fn delete_migrating_vm(
&mut self,
request: impl IntoRequest<DeleteMigratingVmRequest>
) -> Result<Response<Operation>, Status>

Deletes a single MigratingVm.

pub async fn start_migration(
&mut self,
request: impl IntoRequest<StartMigrationRequest>
) -> Result<Response<Operation>, Status>

Starts migration for a VM. Starts the process of uploading data and creating snapshots, in replication cycles scheduled by the policy.

pub async fn resume_migration(
&mut self,
request: impl IntoRequest<ResumeMigrationRequest>
) -> Result<Response<Operation>, Status>

Resumes a migration for a VM. When called on a paused migration, will start the process of uploading data and creating snapshots; when called on a completed cut-over migration, will update the migration to active state and start the process of uploading data and creating snapshots.

pub async fn pause_migration(
&mut self,
request: impl IntoRequest<PauseMigrationRequest>
) -> Result<Response<Operation>, Status>

Pauses a migration for a VM. If cycle tasks are running they will be cancelled, preserving source task data. Further replication cycles will not be triggered while the VM is paused.

pub async fn finalize_migration(
&mut self,
request: impl IntoRequest<FinalizeMigrationRequest>
) -> Result<Response<Operation>, Status>

Marks a migration as completed, deleting migration resources that are no longer being used. Only applicable after cutover is done.

pub async fn create_clone_job(
&mut self,
request: impl IntoRequest<CreateCloneJobRequest>
) -> Result<Response<Operation>, Status>

Initiates a Clone of a specific migrating VM.

pub async fn cancel_clone_job(
&mut self,
request: impl IntoRequest<CancelCloneJobRequest>
) -> Result<Response<Operation>, Status>

Initiates the cancellation of a running clone job.

pub async fn list_clone_jobs(
&mut self,
request: impl IntoRequest<ListCloneJobsRequest>
) -> Result<Response<ListCloneJobsResponse>, Status>

Lists CloneJobs of a given migrating VM.

pub async fn get_clone_job(
&mut self,
request: impl IntoRequest<GetCloneJobRequest>
) -> Result<Response<CloneJob>, Status>

Gets details of a single CloneJob.

pub async fn create_cutover_job(
&mut self,
request: impl IntoRequest<CreateCutoverJobRequest>
) -> Result<Response<Operation>, Status>

Initiates a Cutover of a specific migrating VM. The returned LRO is completed when the cutover job resource is created and the job is initiated.

pub async fn cancel_cutover_job(
&mut self,
request: impl IntoRequest<CancelCutoverJobRequest>
) -> Result<Response<Operation>, Status>

Initiates the cancellation of a running cutover job.

pub async fn list_cutover_jobs(
&mut self,
request: impl IntoRequest<ListCutoverJobsRequest>
) -> Result<Response<ListCutoverJobsResponse>, Status>

Lists CutoverJobs of a given migrating VM.

pub async fn get_cutover_job(
&mut self,
request: impl IntoRequest<GetCutoverJobRequest>
) -> Result<Response<CutoverJob>, Status>

Gets details of a single CutoverJob.

pub async fn list_groups(
&mut self,
request: impl IntoRequest<ListGroupsRequest>
) -> Result<Response<ListGroupsResponse>, Status>

Lists Groups in a given project and location.

pub async fn get_group(
&mut self,
request: impl IntoRequest<GetGroupRequest>
) -> Result<Response<Group>, Status>

Gets details of a single Group.

pub async fn create_group(
&mut self,
request: impl IntoRequest<CreateGroupRequest>
) -> Result<Response<Operation>, Status>

Creates a new Group in a given project and location.

pub async fn update_group(
&mut self,
request: impl IntoRequest<UpdateGroupRequest>
) -> Result<Response<Operation>, Status>

Updates the parameters of a single Group.

pub async fn delete_group(
&mut self,
request: impl IntoRequest<DeleteGroupRequest>
) -> Result<Response<Operation>, Status>

Deletes a single Group.

pub async fn add_group_migration(
&mut self,
request: impl IntoRequest<AddGroupMigrationRequest>
) -> Result<Response<Operation>, Status>

Adds a MigratingVm to a Group.

pub async fn remove_group_migration(
&mut self,
request: impl IntoRequest<RemoveGroupMigrationRequest>
) -> Result<Response<Operation>, Status>

Removes a MigratingVm from a Group.

pub async fn list_target_projects(
&mut self,
request: impl IntoRequest<ListTargetProjectsRequest>
) -> Result<Response<ListTargetProjectsResponse>, Status>

Lists TargetProjects in a given project.

NOTE: TargetProject is a global resource; hence the only supported value for location is global.

pub async fn get_target_project(
&mut self,
request: impl IntoRequest<GetTargetProjectRequest>
) -> Result<Response<TargetProject>, Status>

Gets details of a single TargetProject.

NOTE: TargetProject is a global resource; hence the only supported value for location is global.

pub async fn create_target_project(
&mut self,
request: impl IntoRequest<CreateTargetProjectRequest>
) -> Result<Response<Operation>, Status>

Creates a new TargetProject in a given project.

NOTE: TargetProject is a global resource; hence the only supported value for location is global.

pub async fn update_target_project(
&mut self,
request: impl IntoRequest<UpdateTargetProjectRequest>
) -> Result<Response<Operation>, Status>

Updates the parameters of a single TargetProject.

NOTE: TargetProject is a global resource; hence the only supported value for location is global.

pub async fn delete_target_project(
&mut self,
request: impl IntoRequest<DeleteTargetProjectRequest>
) -> Result<Response<Operation>, Status>

Deletes a single TargetProject.

NOTE: TargetProject is a global resource; hence the only supported value for location is global.