Module google_api_proto::google::cloud::vmwareengine::v1

Modules

cluster

Nested message and enum types in Cluster.

hcx

Nested message and enum types in Hcx.

hcx_activation_key

Nested message and enum types in HcxActivationKey.

network_policy

Nested message and enum types in NetworkPolicy.

nsx

Nested message and enum types in Nsx.

peering_route

Nested message and enum types in PeeringRoute.

private_cloud

Nested message and enum types in PrivateCloud.

private_connection

Nested message and enum types in PrivateConnection.

subnet

Nested message and enum types in Subnet.

vcenter

Nested message and enum types in Vcenter.

vmware_engine_client

Generated client implementations.

vmware_engine_network

Nested message and enum types in VmwareEngineNetwork.

Structs

Cluster

A cluster in a private cloud.

CreateClusterRequest

Request message for [VmwareEngine.CreateCluster][google.cloud.vmwareengine.v1.VmwareEngine.CreateCluster]

CreateHcxActivationKeyRequest

Request message for [VmwareEngine.CreateHcxActivationKey][google.cloud.vmwareengine.v1.VmwareEngine.CreateHcxActivationKey]

CreateNetworkPolicyRequest

Request message for [VmwareEngine.CreateNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.CreateNetworkPolicy]

CreatePrivateCloudRequest

Request message for [VmwareEngine.CreatePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.CreatePrivateCloud]

CreatePrivateConnectionRequest

Request message for [VmwareEngine.CreatePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.CreatePrivateConnection]

CreateVmwareEngineNetworkRequest

Request message for [VmwareEngine.CreateVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.CreateVmwareEngineNetwork]

Credentials

Credentials for a private cloud.

DeleteClusterRequest

Request message for [VmwareEngine.DeleteCluster][google.cloud.vmwareengine.v1.VmwareEngine.DeleteCluster]

DeleteNetworkPolicyRequest

Request message for [VmwareEngine.DeleteNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.DeleteNetworkPolicy]

DeletePrivateCloudRequest

Request message for [VmwareEngine.DeletePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.DeletePrivateCloud]

DeletePrivateConnectionRequest

Request message for [VmwareEngine.DeletePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.DeletePrivateConnection]

DeleteVmwareEngineNetworkRequest

Request message for [VmwareEngine.DeleteVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.DeleteVmwareEngineNetwork]

GetClusterRequest

Request message for [VmwareEngine.GetCluster][google.cloud.vmwareengine.v1.VmwareEngine.GetCluster]

GetHcxActivationKeyRequest

Request message for [VmwareEngine.GetHcxActivationKeys][]

GetNetworkPolicyRequest

Request message for [VmwareEngine.GetNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.GetNetworkPolicy]

GetNodeTypeRequest

Request message for [VmwareEngine.GetNodeType][google.cloud.vmwareengine.v1.VmwareEngine.GetNodeType]

GetPrivateCloudRequest

Request message for [VmwareEngine.GetPrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.GetPrivateCloud]

GetPrivateConnectionRequest

Request message for [VmwareEngine.GetPrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.GetPrivateConnection]

GetSubnetRequest

Request message for [VmwareEngine.GetSubnet][google.cloud.vmwareengine.v1.VmwareEngine.GetSubnet]

GetVmwareEngineNetworkRequest

Request message for [VmwareEngine.GetVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.GetVmwareEngineNetwork]

Hcx

Details about a HCX Cloud Manager appliance.

HcxActivationKey

HCX activation key. A default key is created during private cloud provisioning, but this behavior is subject to change and you should always verify active keys. Use [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys] to retrieve existing keys and [VmwareEngine.CreateHcxActivationKey][google.cloud.vmwareengine.v1.VmwareEngine.CreateHcxActivationKey] to create new ones.

ListClustersRequest

Request message for [VmwareEngine.ListClusters][google.cloud.vmwareengine.v1.VmwareEngine.ListClusters]

ListClustersResponse

Response message for [VmwareEngine.ListClusters][google.cloud.vmwareengine.v1.VmwareEngine.ListClusters]

ListHcxActivationKeysRequest

Request message for [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys]

ListHcxActivationKeysResponse

Response message for [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys]

ListNetworkPoliciesRequest

Request message for [VmwareEngine.ListNetworkPolicies][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPolicies]

ListNetworkPoliciesResponse

Response message for [VmwareEngine.ListNetworkPolicies][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPolicies]

ListNodeTypesRequest

Request message for [VmwareEngine.ListNodeTypes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodeTypes]

ListNodeTypesResponse

Response message for [VmwareEngine.ListNodeTypes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodeTypes]

ListPrivateCloudsRequest

Request message for [VmwareEngine.ListPrivateClouds][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateClouds]

ListPrivateCloudsResponse

Response message for [VmwareEngine.ListPrivateClouds][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateClouds]

ListPrivateConnectionPeeringRoutesRequest

Request message for [VmwareEngine.ListPrivateConnectionPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnectionPeeringRoutes]

ListPrivateConnectionPeeringRoutesResponse

Response message for [VmwareEngine.ListPrivateConnectionPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnectionPeeringRoutes]

ListPrivateConnectionsRequest

Request message for [VmwareEngine.ListPrivateConnections][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnections]

ListPrivateConnectionsResponse

Response message for [VmwareEngine.ListPrivateConnections][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnections]

ListSubnetsRequest

Request message for [VmwareEngine.ListSubnets][google.cloud.vmwareengine.v1.VmwareEngine.ListSubnets]

ListSubnetsResponse

Response message for [VmwareEngine.ListSubnets][google.cloud.vmwareengine.v1.VmwareEngine.ListSubnets]

ListVmwareEngineNetworksRequest

Request message for [VmwareEngine.ListVmwareEngineNetworks][google.cloud.vmwareengine.v1.VmwareEngine.ListVmwareEngineNetworks]

ListVmwareEngineNetworksResponse

Response message for [VmwareEngine.ListVmwareEngineNetworks][google.cloud.vmwareengine.v1.VmwareEngine.ListVmwareEngineNetworks]

NetworkConfig

Network configuration in the consumer project with which the peering has to be done.

NetworkPolicy

Represents a network policy resource. Network policies are regional resources. You can use a network policy to enable or disable internet access and external IP access. Network policies are associated with a VMware Engine network, which might span across regions. For a given region, a network policy applies to all private clouds in the VMware Engine network associated with the policy.

NodeType

Describes node type.

NodeTypeConfig

Information about the type and number of nodes associated with the cluster.

Nsx

Details about a NSX Manager appliance.

OperationMetadata

Represents the metadata of the long-running operation.

PeeringRoute

Exchanged network peering route.

PrivateCloud

Represents a private cloud resource. Private clouds are zonal resources.

PrivateConnection

Private connection resource that provides connectivity for VMware Engine private clouds.

ResetNsxCredentialsRequest

Request message for [VmwareEngine.ResetNsxCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ResetNsxCredentials]

ResetVcenterCredentialsRequest

Request message for [VmwareEngine.ResetVcenterCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ResetVcenterCredentials]

ShowNsxCredentialsRequest

Request message for [VmwareEngine.ShowNsxCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ShowNsxCredentials]

ShowVcenterCredentialsRequest

Request message for [VmwareEngine.ShowVcenterCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ShowVcenterCredentials]

Subnet

Subnet in a private cloud. Either management subnets (such as vMotion) that are read-only, or userDefined, which can also be updated.

UndeletePrivateCloudRequest

Request message for [VmwareEngine.UndeletePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.UndeletePrivateCloud]

UpdateClusterRequest

Request message for [VmwareEngine.UpdateCluster][google.cloud.vmwareengine.v1.VmwareEngine.UpdateCluster]

UpdateNetworkPolicyRequest

Request message for [VmwareEngine.UpdateNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.UpdateNetworkPolicy]

UpdatePrivateCloudRequest

Request message for [VmwareEngine.UpdatePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.UpdatePrivateCloud]

UpdatePrivateConnectionRequest

Request message for [VmwareEngine.UpdatePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.UpdatePrivateConnection]

UpdateSubnetRequest

Request message for [VmwareEngine.UpdateSubnet][google.cloud.vmwareengine.v1.VmwareEngine.UpdateSubnet]

UpdateVmwareEngineNetworkRequest

Request message for [VmwareEngine.UpdateVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.UpdateVmwareEngineNetwork]

Vcenter

Details about a vCenter Server management appliance.

VmwareEngineNetwork

VMware Engine network resource that provides connectivity for VMware Engine private clouds.