Module google_api_proto::google::cloud::vmwareengine::v1

Modules

cluster

Nested message and enum types in Cluster.

dns_forwarding

Nested message and enum types in DnsForwarding.

external_access_rule

Nested message and enum types in ExternalAccessRule.

external_address

Nested message and enum types in ExternalAddress.

hcx

Nested message and enum types in Hcx.

hcx_activation_key

Nested message and enum types in HcxActivationKey.

location_metadata

Nested message and enum types in LocationMetadata.

logging_server

Nested message and enum types in LoggingServer.

management_dns_zone_binding

Nested message and enum types in ManagementDnsZoneBinding.

network_peering

Nested message and enum types in NetworkPeering.

network_policy

Nested message and enum types in NetworkPolicy.

node

Nested message and enum types in Node.

node_type

Nested message and enum types in NodeType.

nsx

Nested message and enum types in Nsx.

peering_route

Nested message and enum types in PeeringRoute.

principal

Nested message and enum types in Principal.

private_cloud

Nested message and enum types in PrivateCloud.

private_connection

Nested message and enum types in PrivateConnection.

subnet

Nested message and enum types in Subnet.

vcenter

Nested message and enum types in Vcenter.

vmware_engine_client

Generated client implementations.

vmware_engine_network

Nested message and enum types in VmwareEngineNetwork.

Structs

Cluster

A cluster in a private cloud.

CreateClusterRequest

Request message for [VmwareEngine.CreateCluster][google.cloud.vmwareengine.v1.VmwareEngine.CreateCluster]

CreateExternalAccessRuleRequest

Request message for [VmwareEngine.CreateExternalAccessRule][google.cloud.vmwareengine.v1.VmwareEngine.CreateExternalAccessRule]

CreateExternalAddressRequest

Request message for [VmwareEngine.CreateExternalAddress][google.cloud.vmwareengine.v1.VmwareEngine.CreateExternalAddress]

CreateHcxActivationKeyRequest

Request message for [VmwareEngine.CreateHcxActivationKey][google.cloud.vmwareengine.v1.VmwareEngine.CreateHcxActivationKey]

CreateLoggingServerRequest

Request message for [VmwareEngine.CreateLoggingServer][google.cloud.vmwareengine.v1.VmwareEngine.CreateLoggingServer]

CreateManagementDnsZoneBindingRequest

Request message for [VmwareEngine.CreateManagementDnsZoneBindings][]

CreateNetworkPeeringRequest

Request message for [VmwareEngine.CreateNetworkPeering][google.cloud.vmwareengine.v1.VmwareEngine.CreateNetworkPeering]

CreateNetworkPolicyRequest

Request message for [VmwareEngine.CreateNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.CreateNetworkPolicy]

CreatePrivateCloudRequest

Request message for [VmwareEngine.CreatePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.CreatePrivateCloud]

CreatePrivateConnectionRequest

Request message for [VmwareEngine.CreatePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.CreatePrivateConnection]

CreateVmwareEngineNetworkRequest

Request message for [VmwareEngine.CreateVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.CreateVmwareEngineNetwork]

Credentials

Credentials for a private cloud.

DeleteClusterRequest

Request message for [VmwareEngine.DeleteCluster][google.cloud.vmwareengine.v1.VmwareEngine.DeleteCluster]

DeleteExternalAccessRuleRequest

Request message for [VmwareEngine.DeleteExternalAccessRule][google.cloud.vmwareengine.v1.VmwareEngine.DeleteExternalAccessRule]

DeleteExternalAddressRequest

Request message for [VmwareEngine.DeleteExternalAddress][google.cloud.vmwareengine.v1.VmwareEngine.DeleteExternalAddress]

DeleteLoggingServerRequest

Request message for [VmwareEngine.DeleteLoggingServer][google.cloud.vmwareengine.v1.VmwareEngine.DeleteLoggingServer]

DeleteManagementDnsZoneBindingRequest

Request message for [VmwareEngine.DeleteManagementDnsZoneBinding][google.cloud.vmwareengine.v1.VmwareEngine.DeleteManagementDnsZoneBinding]

DeleteNetworkPeeringRequest

Request message for [VmwareEngine.DeleteNetworkPeering][google.cloud.vmwareengine.v1.VmwareEngine.DeleteNetworkPeering]

DeleteNetworkPolicyRequest

Request message for [VmwareEngine.DeleteNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.DeleteNetworkPolicy]

DeletePrivateCloudRequest

Request message for [VmwareEngine.DeletePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.DeletePrivateCloud]

DeletePrivateConnectionRequest

Request message for [VmwareEngine.DeletePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.DeletePrivateConnection]

DeleteVmwareEngineNetworkRequest

Request message for [VmwareEngine.DeleteVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.DeleteVmwareEngineNetwork]

DnsBindPermission

DnsBindPermission resource that contains the accounts having the consumer DNS bind permission on the corresponding intranet VPC of the consumer project.

DnsForwarding

DNS forwarding config. This config defines a list of domain to name server mappings, and is attached to the private cloud for custom domain resolution.

ExternalAccessRule

External access firewall rules for filtering incoming traffic destined to ExternalAddress resources.

ExternalAddress

Represents an allocated external IP address and its corresponding internal IP address in a private cloud.

FetchNetworkPolicyExternalAddressesRequest

Request message for [VmwareEngine.FetchNetworkPolicyExternalAddresses][google.cloud.vmwareengine.v1.VmwareEngine.FetchNetworkPolicyExternalAddresses]

FetchNetworkPolicyExternalAddressesResponse

Response message for [VmwareEngine.FetchNetworkPolicyExternalAddresses][google.cloud.vmwareengine.v1.VmwareEngine.FetchNetworkPolicyExternalAddresses]

GetClusterRequest

Request message for [VmwareEngine.GetCluster][google.cloud.vmwareengine.v1.VmwareEngine.GetCluster]

GetDnsBindPermissionRequest

Request message for [VmwareEngine.GetDnsBindPermission][google.cloud.vmwareengine.v1.VmwareEngine.GetDnsBindPermission]

GetDnsForwardingRequest

Request message for [VmwareEngine.GetDnsForwarding][google.cloud.vmwareengine.v1.VmwareEngine.GetDnsForwarding]

GetExternalAccessRuleRequest

Request message for [VmwareEngine.GetExternalAccessRule][google.cloud.vmwareengine.v1.VmwareEngine.GetExternalAccessRule]

GetExternalAddressRequest

Request message for [VmwareEngine.GetExternalAddress][google.cloud.vmwareengine.v1.VmwareEngine.GetExternalAddress]

GetHcxActivationKeyRequest

Request message for [VmwareEngine.GetHcxActivationKeys][]

GetLoggingServerRequest

Request message for [VmwareEngine.GetLoggingServer][google.cloud.vmwareengine.v1.VmwareEngine.GetLoggingServer]

GetManagementDnsZoneBindingRequest

Request message for [VmwareEngine.GetManagementDnsZoneBinding][google.cloud.vmwareengine.v1.VmwareEngine.GetManagementDnsZoneBinding]

GetNetworkPeeringRequest

Request message for [VmwareEngine.GetNetworkPeering][google.cloud.vmwareengine.v1.VmwareEngine.GetNetworkPeering]

GetNetworkPolicyRequest

Request message for [VmwareEngine.GetNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.GetNetworkPolicy]

GetNodeRequest

Request message for [VmwareEngine.GetNode][google.cloud.vmwareengine.v1.VmwareEngine.GetNode]

GetNodeTypeRequest

Request message for [VmwareEngine.GetNodeType][google.cloud.vmwareengine.v1.VmwareEngine.GetNodeType]

GetPrivateCloudRequest

Request message for [VmwareEngine.GetPrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.GetPrivateCloud]

GetPrivateConnectionRequest

Request message for [VmwareEngine.GetPrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.GetPrivateConnection]

GetSubnetRequest

Request message for [VmwareEngine.GetSubnet][google.cloud.vmwareengine.v1.VmwareEngine.GetSubnet]

GetVmwareEngineNetworkRequest

Request message for [VmwareEngine.GetVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.GetVmwareEngineNetwork]

GrantDnsBindPermissionRequest

Request message for [VmwareEngine.GrantDnsBindPermission][google.cloud.vmwareengine.v1.VmwareEngine.GrantDnsBindPermission]

Hcx

Details about a HCX Cloud Manager appliance.

HcxActivationKey

HCX activation key. A default key is created during private cloud provisioning, but this behavior is subject to change and you should always verify active keys. Use [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys] to retrieve existing keys and [VmwareEngine.CreateHcxActivationKey][google.cloud.vmwareengine.v1.VmwareEngine.CreateHcxActivationKey] to create new ones.

ListClustersRequest

Request message for [VmwareEngine.ListClusters][google.cloud.vmwareengine.v1.VmwareEngine.ListClusters]

ListClustersResponse

Response message for [VmwareEngine.ListClusters][google.cloud.vmwareengine.v1.VmwareEngine.ListClusters]

ListExternalAccessRulesRequest

Request message for [VmwareEngine.ListExternalAccessRules][google.cloud.vmwareengine.v1.VmwareEngine.ListExternalAccessRules]

ListExternalAccessRulesResponse

Response message for [VmwareEngine.ListExternalAccessRules][google.cloud.vmwareengine.v1.VmwareEngine.ListExternalAccessRules]

ListExternalAddressesRequest

Request message for [VmwareEngine.ListExternalAddresses][google.cloud.vmwareengine.v1.VmwareEngine.ListExternalAddresses]

ListExternalAddressesResponse

Response message for [VmwareEngine.ListExternalAddresses][google.cloud.vmwareengine.v1.VmwareEngine.ListExternalAddresses]

ListHcxActivationKeysRequest

Request message for [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys]

ListHcxActivationKeysResponse

Response message for [VmwareEngine.ListHcxActivationKeys][google.cloud.vmwareengine.v1.VmwareEngine.ListHcxActivationKeys]

ListLoggingServersRequest

Request message for [VmwareEngine.ListLoggingServers][google.cloud.vmwareengine.v1.VmwareEngine.ListLoggingServers]

ListLoggingServersResponse

Response message for [VmwareEngine.ListLoggingServers][google.cloud.vmwareengine.v1.VmwareEngine.ListLoggingServers]

ListManagementDnsZoneBindingsRequest

Request message for [VmwareEngine.ListManagementDnsZoneBindings][google.cloud.vmwareengine.v1.VmwareEngine.ListManagementDnsZoneBindings]

ListManagementDnsZoneBindingsResponse

Response message for [VmwareEngine.ListManagementDnsZoneBindings][google.cloud.vmwareengine.v1.VmwareEngine.ListManagementDnsZoneBindings]

ListNetworkPeeringsRequest

Request message for [VmwareEngine.ListNetworkPeerings][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPeerings]

ListNetworkPeeringsResponse

Response message for [VmwareEngine.ListNetworkPeerings][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPeerings]

ListNetworkPoliciesRequest

Request message for [VmwareEngine.ListNetworkPolicies][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPolicies]

ListNetworkPoliciesResponse

Response message for [VmwareEngine.ListNetworkPolicies][google.cloud.vmwareengine.v1.VmwareEngine.ListNetworkPolicies]

ListNodeTypesRequest

Request message for [VmwareEngine.ListNodeTypes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodeTypes]

ListNodeTypesResponse

Response message for [VmwareEngine.ListNodeTypes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodeTypes]

ListNodesRequest

Request message for [VmwareEngine.ListNodes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodes]

ListNodesResponse

Response message for [VmwareEngine.ListNodes][google.cloud.vmwareengine.v1.VmwareEngine.ListNodes]

ListPeeringRoutesRequest

Request message for [VmwareEngine.ListPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPeeringRoutes]

ListPeeringRoutesResponse

Response message for [VmwareEngine.ListPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPeeringRoutes]

ListPrivateCloudsRequest

Request message for [VmwareEngine.ListPrivateClouds][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateClouds]

ListPrivateCloudsResponse

Response message for [VmwareEngine.ListPrivateClouds][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateClouds]

ListPrivateConnectionPeeringRoutesRequest

Request message for [VmwareEngine.ListPrivateConnectionPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnectionPeeringRoutes]

ListPrivateConnectionPeeringRoutesResponse

Response message for [VmwareEngine.ListPrivateConnectionPeeringRoutes][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnectionPeeringRoutes]

ListPrivateConnectionsRequest

Request message for [VmwareEngine.ListPrivateConnections][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnections]

ListPrivateConnectionsResponse

Response message for [VmwareEngine.ListPrivateConnections][google.cloud.vmwareengine.v1.VmwareEngine.ListPrivateConnections]

ListSubnetsRequest

Request message for [VmwareEngine.ListSubnets][google.cloud.vmwareengine.v1.VmwareEngine.ListSubnets]

ListSubnetsResponse

Response message for [VmwareEngine.ListSubnets][google.cloud.vmwareengine.v1.VmwareEngine.ListSubnets]

ListVmwareEngineNetworksRequest

Request message for [VmwareEngine.ListVmwareEngineNetworks][google.cloud.vmwareengine.v1.VmwareEngine.ListVmwareEngineNetworks]

ListVmwareEngineNetworksResponse

Response message for [VmwareEngine.ListVmwareEngineNetworks][google.cloud.vmwareengine.v1.VmwareEngine.ListVmwareEngineNetworks]

LocationMetadata

VmwareEngine specific metadata for the given [google.cloud.location.Location][google.cloud.location.Location]. It is returned as a content of the google.cloud.location.Location.metadata field.

LoggingServer

Logging server to receive vCenter or ESXi logs.

ManagementDnsZoneBinding

Represents a binding between a network and the management DNS zone. A management DNS zone is the Cloud DNS cross-project binding zone that VMware Engine creates for each private cloud. It contains FQDNs and corresponding IP addresses for the private cloud’s ESXi hosts and management VM appliances like vCenter and NSX Manager.

NetworkConfig

Network configuration in the consumer project with which the peering has to be done.

NetworkPeering

Details of a network peering.

NetworkPolicy

Represents a network policy resource. Network policies are regional resources. You can use a network policy to enable or disable internet access and external IP access. Network policies are associated with a VMware Engine network, which might span across regions. For a given region, a network policy applies to all private clouds in the VMware Engine network associated with the policy.

Node

Node in a cluster.

NodeType

Describes node type.

NodeTypeConfig

Information about the type and number of nodes associated with the cluster.

Nsx

Details about a NSX Manager appliance.

OperationMetadata

Represents the metadata of the long-running operation.

PeeringRoute

Exchanged network peering route.

Principal

Users/Service accounts which have access for DNS binding on the intranet VPC corresponding to the consumer project.

PrivateCloud

Represents a private cloud resource. Private clouds of type STANDARD and TIME_LIMITED are zonal resources, STRETCHED private clouds are regional.

PrivateConnection

Private connection resource that provides connectivity for VMware Engine private clouds.

RepairManagementDnsZoneBindingRequest

Request message for [VmwareEngine.RepairManagementDnsZoneBindings][]

ResetNsxCredentialsRequest

Request message for [VmwareEngine.ResetNsxCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ResetNsxCredentials]

ResetVcenterCredentialsRequest

Request message for [VmwareEngine.ResetVcenterCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ResetVcenterCredentials]

RevokeDnsBindPermissionRequest

Request message for [VmwareEngine.RevokeDnsBindPermission][google.cloud.vmwareengine.v1.VmwareEngine.RevokeDnsBindPermission]

ShowNsxCredentialsRequest

Request message for [VmwareEngine.ShowNsxCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ShowNsxCredentials]

ShowVcenterCredentialsRequest

Request message for [VmwareEngine.ShowVcenterCredentials][google.cloud.vmwareengine.v1.VmwareEngine.ShowVcenterCredentials]

StretchedClusterConfig

Configuration of a stretched cluster.

Subnet

Subnet in a private cloud. Either management subnets (such as vMotion) that are read-only, or userDefined, which can also be updated.

UndeletePrivateCloudRequest

Request message for [VmwareEngine.UndeletePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.UndeletePrivateCloud]

UpdateClusterRequest

Request message for [VmwareEngine.UpdateCluster][google.cloud.vmwareengine.v1.VmwareEngine.UpdateCluster]

UpdateDnsForwardingRequest

Request message for [VmwareEngine.UpdateDnsForwarding][google.cloud.vmwareengine.v1.VmwareEngine.UpdateDnsForwarding]

UpdateExternalAccessRuleRequest

Request message for [VmwareEngine.UpdateExternalAccessRule][google.cloud.vmwareengine.v1.VmwareEngine.UpdateExternalAccessRule]

UpdateExternalAddressRequest

Request message for [VmwareEngine.UpdateExternalAddress][google.cloud.vmwareengine.v1.VmwareEngine.UpdateExternalAddress]

UpdateLoggingServerRequest

Request message for [VmwareEngine.UpdateLoggingServer][google.cloud.vmwareengine.v1.VmwareEngine.UpdateLoggingServer]

UpdateManagementDnsZoneBindingRequest

Request message for [VmwareEngine.UpdateManagementDnsZoneBinding][google.cloud.vmwareengine.v1.VmwareEngine.UpdateManagementDnsZoneBinding]

UpdateNetworkPeeringRequest

Request message for [VmwareEngine.UpdateNetworkPeering][google.cloud.vmwareengine.v1.VmwareEngine.UpdateNetworkPeering]

UpdateNetworkPolicyRequest

Request message for [VmwareEngine.UpdateNetworkPolicy][google.cloud.vmwareengine.v1.VmwareEngine.UpdateNetworkPolicy]

UpdatePrivateCloudRequest

Request message for [VmwareEngine.UpdatePrivateCloud][google.cloud.vmwareengine.v1.VmwareEngine.UpdatePrivateCloud]

UpdatePrivateConnectionRequest

Request message for [VmwareEngine.UpdatePrivateConnection][google.cloud.vmwareengine.v1.VmwareEngine.UpdatePrivateConnection]

UpdateSubnetRequest

Request message for [VmwareEngine.UpdateSubnet][google.cloud.vmwareengine.v1.VmwareEngine.UpdateSubnet]

UpdateVmwareEngineNetworkRequest

Request message for [VmwareEngine.UpdateVmwareEngineNetwork][google.cloud.vmwareengine.v1.VmwareEngine.UpdateVmwareEngineNetwork]

Vcenter

Details about a vCenter Server management appliance.

VmwareEngineNetwork

VMware Engine network resource that provides connectivity for VMware Engine private clouds.