Struct google_api_proto::google::cloud::vmwareengine::v1::vmware_engine_client::VmwareEngineClient

pub struct VmwareEngineClient<T> { /* private fields */ }

VMwareEngine manages VMware’s private clusters in the Cloud.

Implementations

impl<T> VmwareEngineClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> VmwareEngineClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_private_clouds( &mut self, request: impl IntoRequest<ListPrivateCloudsRequest> ) -> Result<Response<ListPrivateCloudsResponse>, Status>

Lists PrivateCloud resources in a given project and location.

pub async fn get_private_cloud( &mut self, request: impl IntoRequest<GetPrivateCloudRequest> ) -> Result<Response<PrivateCloud>, Status>

Retrieves a PrivateCloud resource by its resource name.

pub async fn create_private_cloud( &mut self, request: impl IntoRequest<CreatePrivateCloudRequest> ) -> Result<Response<Operation>, Status>

Creates a new PrivateCloud resource in a given project and location. Private clouds can only be created in zones, regional private clouds are not supported.

Creating a private cloud also creates a management cluster for that private cloud.

pub async fn update_private_cloud( &mut self, request: impl IntoRequest<UpdatePrivateCloudRequest> ) -> Result<Response<Operation>, Status>

Modifies a PrivateCloud resource. Only the following fields can be updated: description. Only fields specified in updateMask are applied.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_private_cloud( &mut self, request: impl IntoRequest<DeletePrivateCloudRequest> ) -> Result<Response<Operation>, Status>

Schedules a PrivateCloud resource for deletion.

A PrivateCloud resource scheduled for deletion has PrivateCloud.state set to DELETED and expireTime set to the time when deletion is final and can no longer be reversed. The delete operation is marked as done as soon as the PrivateCloud is successfully scheduled for deletion (this also applies when delayHours is set to zero), and the operation is not kept in pending state until PrivateCloud is purged. PrivateCloud can be restored using UndeletePrivateCloud method before the expireTime elapses. When expireTime is reached, deletion is final and all private cloud resources are irreversibly removed and billing stops. During the final removal process, PrivateCloud.state is set to PURGING. PrivateCloud can be polled using standard GET method for the whole period of deletion and purging. It will not be returned only when it is completely purged.

pub async fn undelete_private_cloud( &mut self, request: impl IntoRequest<UndeletePrivateCloudRequest> ) -> Result<Response<Operation>, Status>

Restores a private cloud that was previously scheduled for deletion by DeletePrivateCloud. A PrivateCloud resource scheduled for deletion has PrivateCloud.state set to DELETED and PrivateCloud.expireTime set to the time when deletion can no longer be reversed.

pub async fn list_clusters( &mut self, request: impl IntoRequest<ListClustersRequest> ) -> Result<Response<ListClustersResponse>, Status>

Lists Cluster resources in a given private cloud.

pub async fn get_cluster( &mut self, request: impl IntoRequest<GetClusterRequest> ) -> Result<Response<Cluster>, Status>

Retrieves a Cluster resource by its resource name.

pub async fn create_cluster( &mut self, request: impl IntoRequest<CreateClusterRequest> ) -> Result<Response<Operation>, Status>

Creates a new cluster in a given private cloud. Creating a new cluster provides additional nodes for use in the parent private cloud and requires sufficient node quota.

pub async fn update_cluster( &mut self, request: impl IntoRequest<UpdateClusterRequest> ) -> Result<Response<Operation>, Status>

Modifies a Cluster resource. Only the following fields can be updated: node_type_configs.*.node_count. Only fields specified in updateMask are applied.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_cluster( &mut self, request: impl IntoRequest<DeleteClusterRequest> ) -> Result<Response<Operation>, Status>

Deletes a Cluster resource. To avoid unintended data loss, migrate or gracefully shut down any workloads running on the cluster before deletion. You cannot delete the management cluster of a private cloud using this method.

pub async fn list_subnets( &mut self, request: impl IntoRequest<ListSubnetsRequest> ) -> Result<Response<ListSubnetsResponse>, Status>

Lists subnets in a given private cloud.

pub async fn get_subnet( &mut self, request: impl IntoRequest<GetSubnetRequest> ) -> Result<Response<Subnet>, Status>

Gets details of a single subnet.

pub async fn update_subnet( &mut self, request: impl IntoRequest<UpdateSubnetRequest> ) -> Result<Response<Operation>, Status>

Updates the parameters of a single subnet. Only fields specified in update_mask are applied.

Note: This API is synchronous and always returns a successful google.longrunning.Operation (LRO). The returned LRO will only have done and response fields.

pub async fn list_node_types( &mut self, request: impl IntoRequest<ListNodeTypesRequest> ) -> Result<Response<ListNodeTypesResponse>, Status>

Lists node types

pub async fn get_node_type( &mut self, request: impl IntoRequest<GetNodeTypeRequest> ) -> Result<Response<NodeType>, Status>

Gets details of a single NodeType.

pub async fn show_nsx_credentials( &mut self, request: impl IntoRequest<ShowNsxCredentialsRequest> ) -> Result<Response<Credentials>, Status>

Gets details of credentials for NSX appliance.

pub async fn show_vcenter_credentials( &mut self, request: impl IntoRequest<ShowVcenterCredentialsRequest> ) -> Result<Response<Credentials>, Status>

Gets details of credentials for Vcenter appliance.

pub async fn reset_nsx_credentials( &mut self, request: impl IntoRequest<ResetNsxCredentialsRequest> ) -> Result<Response<Operation>, Status>

Resets credentials of the NSX appliance.

pub async fn reset_vcenter_credentials( &mut self, request: impl IntoRequest<ResetVcenterCredentialsRequest> ) -> Result<Response<Operation>, Status>

Resets credentials of the Vcenter appliance.

pub async fn create_hcx_activation_key( &mut self, request: impl IntoRequest<CreateHcxActivationKeyRequest> ) -> Result<Response<Operation>, Status>

Creates a new HCX activation key in a given private cloud.

pub async fn list_hcx_activation_keys( &mut self, request: impl IntoRequest<ListHcxActivationKeysRequest> ) -> Result<Response<ListHcxActivationKeysResponse>, Status>

Lists HcxActivationKey resources in a given private cloud.

pub async fn get_hcx_activation_key( &mut self, request: impl IntoRequest<GetHcxActivationKeyRequest> ) -> Result<Response<HcxActivationKey>, Status>

Retrieves a HcxActivationKey resource by its resource name.

pub async fn get_network_policy( &mut self, request: impl IntoRequest<GetNetworkPolicyRequest> ) -> Result<Response<NetworkPolicy>, Status>

Retrieves a NetworkPolicy resource by its resource name.

pub async fn list_network_policies( &mut self, request: impl IntoRequest<ListNetworkPoliciesRequest> ) -> Result<Response<ListNetworkPoliciesResponse>, Status>

Lists NetworkPolicy resources in a specified project and location.

pub async fn create_network_policy( &mut self, request: impl IntoRequest<CreateNetworkPolicyRequest> ) -> Result<Response<Operation>, Status>

Creates a new network policy in a given VMware Engine network of a project and location (region). A new network policy cannot be created if another network policy already exists in the same scope.

pub async fn update_network_policy( &mut self, request: impl IntoRequest<UpdateNetworkPolicyRequest> ) -> Result<Response<Operation>, Status>

Modifies a NetworkPolicy resource. Only the following fields can be updated: internet_access, external_ip, edge_services_cidr. Only fields specified in updateMask are applied. When updating a network policy, the external IP network service can only be disabled if there are no external IP addresses present in the scope of the policy. Also, a NetworkService cannot be updated when NetworkService.state is set to RECONCILING.

During operation processing, the resource is temporarily in the ACTIVE state before the operation fully completes. For that period of time, you can’t update the resource. Use the operation status to determine when the processing fully completes.

pub async fn delete_network_policy( &mut self, request: impl IntoRequest<DeleteNetworkPolicyRequest> ) -> Result<Response<Operation>, Status>

Deletes a NetworkPolicy resource. A network policy cannot be deleted when NetworkService.state is set to RECONCILING for either its external IP or internet access service.

pub async fn create_vmware_engine_network( &mut self, request: impl IntoRequest<CreateVmwareEngineNetworkRequest> ) -> Result<Response<Operation>, Status>

Creates a new VMware Engine network that can be used by a private cloud.

pub async fn update_vmware_engine_network( &mut self, request: impl IntoRequest<UpdateVmwareEngineNetworkRequest> ) -> Result<Response<Operation>, Status>

Modifies a VMware Engine network resource. Only the following fields can be updated: description. Only fields specified in updateMask are applied.

pub async fn delete_vmware_engine_network( &mut self, request: impl IntoRequest<DeleteVmwareEngineNetworkRequest> ) -> Result<Response<Operation>, Status>

Deletes a VmwareEngineNetwork resource. You can only delete a VMware Engine network after all resources that refer to it are deleted. For example, a private cloud, a network peering, and a network policy can all refer to the same VMware Engine network.

pub async fn get_vmware_engine_network( &mut self, request: impl IntoRequest<GetVmwareEngineNetworkRequest> ) -> Result<Response<VmwareEngineNetwork>, Status>

Retrieves a VmwareEngineNetwork resource by its resource name. The resource contains details of the VMware Engine network, such as its VMware Engine network type, peered networks in a service project, and state (for example, CREATING, ACTIVE, DELETING).

pub async fn list_vmware_engine_networks( &mut self, request: impl IntoRequest<ListVmwareEngineNetworksRequest> ) -> Result<Response<ListVmwareEngineNetworksResponse>, Status>

Lists VmwareEngineNetwork resources in a given project and location.

pub async fn create_private_connection( &mut self, request: impl IntoRequest<CreatePrivateConnectionRequest> ) -> Result<Response<Operation>, Status>

Creates a new private connection that can be used for accessing private Clouds.

pub async fn get_private_connection( &mut self, request: impl IntoRequest<GetPrivateConnectionRequest> ) -> Result<Response<PrivateConnection>, Status>

Retrieves a PrivateConnection resource by its resource name. The resource contains details of the private connection, such as connected network, routing mode and state.

pub async fn list_private_connections( &mut self, request: impl IntoRequest<ListPrivateConnectionsRequest> ) -> Result<Response<ListPrivateConnectionsResponse>, Status>

Lists PrivateConnection resources in a given project and location.

pub async fn update_private_connection( &mut self, request: impl IntoRequest<UpdatePrivateConnectionRequest> ) -> Result<Response<Operation>, Status>

Modifies a PrivateConnection resource. Only description and routing_mode fields can be updated. Only fields specified in updateMask are applied.

pub async fn delete_private_connection( &mut self, request: impl IntoRequest<DeletePrivateConnectionRequest> ) -> Result<Response<Operation>, Status>

Deletes a PrivateConnection resource. When a private connection is deleted for a VMware Engine network, the connected network becomes inaccessible to that VMware Engine network.

pub async fn list_private_connection_peering_routes( &mut self, request: impl IntoRequest<ListPrivateConnectionPeeringRoutesRequest> ) -> Result<Response<ListPrivateConnectionPeeringRoutesResponse>, Status>

Lists the private connection routes exchanged over a peering connection.