Enum google_api_proto::google::cloud::websecurityscanner::v1::scan_config_error::Code

#[repr(i32)]
pub enum Code {

    Unspecified,
    InternalError,
    AppengineApiBackendError,
    AppengineApiNotAccessible,
    AppengineDefaultHostMissing,
    CannotUseGoogleComAccount,
    CannotUseOwnerAccount,
    ComputeApiBackendError,
    ComputeApiNotAccessible,
    CustomLoginUrlDoesNotBelongToCurrentProject,
    CustomLoginUrlMalformed,
    CustomLoginUrlMappedToNonRoutableAddress,
    CustomLoginUrlMappedToUnreservedAddress,
    CustomLoginUrlHasNonRoutableIpAddress,
    CustomLoginUrlHasUnreservedIpAddress,
    DuplicateScanName,
    InvalidFieldValue,
    FailedToAuthenticateToTarget,
    FindingTypeUnspecified,
    ForbiddenToScanCompute,
    ForbiddenUpdateToManagedScan,
    MalformedFilter,
    MalformedResourceName,
    ProjectInactive,
    RequiredField,
    ResourceNameInconsistent,
    ScanAlreadyRunning,
    ScanNotRunning,
    SeedUrlDoesNotBelongToCurrentProject,
    SeedUrlMalformed,
    SeedUrlMappedToNonRoutableAddress,
    SeedUrlMappedToUnreservedAddress,
    SeedUrlHasNonRoutableIpAddress,
    SeedUrlHasUnreservedIpAddress,
    ServiceAccountNotConfigured,
    TooManyScans,
    UnableToResolveProjectInfo,
    UnsupportedBlacklistPatternFormat,
    UnsupportedFilter,
    UnsupportedFindingType,
    UnsupportedUrlScheme,

}

Output only. Defines an error reason code. Next id: 44

Variants

Unspecified

There is no error.

InternalError

Indicates an internal server error. Please DO NOT USE THIS ERROR CODE unless the root cause is truly unknown.

AppengineApiBackendError

One of the seed URLs is an App Engine URL but we cannot validate the scan settings due to an App Engine API backend error.

AppengineApiNotAccessible

One of the seed URLs is an App Engine URL but we cannot access the App Engine API to validate scan settings.

AppengineDefaultHostMissing

One of the seed URLs is an App Engine URL but the Default Host of the App Engine is not set.

CannotUseGoogleComAccount

Google corporate accounts can not be used for scanning.

CannotUseOwnerAccount

The account of the scan creator can not be used for scanning.

ComputeApiBackendError

This scan targets Compute Engine, but we cannot validate scan settings due to a Compute Engine API backend error.

ComputeApiNotAccessible

This scan targets Compute Engine, but we cannot access the Compute Engine API to validate the scan settings.

CustomLoginUrlDoesNotBelongToCurrentProject

The Custom Login URL does not belong to the current project.

CustomLoginUrlMalformed

The Custom Login URL is malformed (can not be parsed).

CustomLoginUrlMappedToNonRoutableAddress

The Custom Login URL is mapped to a non-routable IP address in DNS.

CustomLoginUrlMappedToUnreservedAddress

The Custom Login URL is mapped to an IP address which is not reserved for the current project.

CustomLoginUrlHasNonRoutableIpAddress

The Custom Login URL has a non-routable IP address.

CustomLoginUrlHasUnreservedIpAddress

The Custom Login URL has an IP address which is not reserved for the current project.

DuplicateScanName

Another scan with the same name (case-sensitive) already exists.

InvalidFieldValue

A field is set to an invalid value.

FailedToAuthenticateToTarget

There was an error trying to authenticate to the scan target.

FindingTypeUnspecified

Finding type value is not specified in the list findings request.

ForbiddenToScanCompute

Scan targets Compute Engine, yet current project was not whitelisted for Google Compute Engine Scanning Alpha access.

ForbiddenUpdateToManagedScan

User tries to update managed scan

MalformedFilter

The supplied filter is malformed. For example, it can not be parsed, does not have a filter type in expression, or the same filter type appears more than once.

MalformedResourceName

The supplied resource name is malformed (can not be parsed).

ProjectInactive

The current project is not in an active state.

RequiredField

A required field is not set.

ResourceNameInconsistent

Project id, scanconfig id, scanrun id, or finding id are not consistent with each other in resource name.

ScanAlreadyRunning

The scan being requested to start is already running.

ScanNotRunning

The scan that was requested to be stopped is not running.

SeedUrlDoesNotBelongToCurrentProject

One of the seed URLs does not belong to the current project.

SeedUrlMalformed

One of the seed URLs is malformed (can not be parsed).

SeedUrlMappedToNonRoutableAddress

One of the seed URLs is mapped to a non-routable IP address in DNS.

SeedUrlMappedToUnreservedAddress

One of the seed URLs is mapped to an IP address which is not reserved for the current project.

SeedUrlHasNonRoutableIpAddress

One of the seed URLs has on-routable IP address.

SeedUrlHasUnreservedIpAddress

One of the seed URLs has an IP address that is not reserved for the current project.

ServiceAccountNotConfigured

The Web Security Scanner service account is not configured under the project.

TooManyScans

A project has reached the maximum number of scans.

UnableToResolveProjectInfo

Resolving the details of the current project fails.

UnsupportedBlacklistPatternFormat

One or more blacklist patterns were in the wrong format.

UnsupportedFilter

The supplied filter is not supported.

UnsupportedFindingType

The supplied finding type is not supported. For example, we do not provide findings of the given finding type.

UnsupportedUrlScheme

The URL scheme of one or more of the supplied URLs is not supported.

Implementations

impl Code

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of Code.

pub fn from_i32(value: i32) -> Option<Code>

Converts an i32 to a Code, or None if value is not a valid variant.

impl Code

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.