Module google_api_proto::google::cloud::workstations::v1beta

Modules

generate_access_token_request

Nested message and enum types in GenerateAccessTokenRequest.

workstation

Nested message and enum types in Workstation.

workstation_cluster

Nested message and enum types in WorkstationCluster.

workstation_config

Nested message and enum types in WorkstationConfig.

workstations_client

Generated client implementations.

Structs

CreateWorkstationClusterRequest

Message for creating a CreateWorkstationCluster.

CreateWorkstationConfigRequest

Message for creating a CreateWorkstationConfig.

CreateWorkstationRequest

Message for creating a CreateWorkstation.

DeleteWorkstationClusterRequest

Message for deleting a workstation cluster.

DeleteWorkstationConfigRequest

Message for deleting a workstation configuration.

DeleteWorkstationRequest

Request message for DeleteWorkstation.

GenerateAccessTokenRequest

Request message for GenerateAccessToken.

GenerateAccessTokenResponse

Response message for GenerateAccessToken.

GetWorkstationClusterRequest

Request message for GetWorkstationCluster.

GetWorkstationConfigRequest

Request message for GetWorkstationConfig.

GetWorkstationRequest

Request message for GetWorkstation.

ListUsableWorkstationConfigsRequest

Request message for ListUsableWorkstationConfigs.

ListUsableWorkstationConfigsResponse

Response message for ListUsableWorkstationConfigs.

ListUsableWorkstationsRequest

Request message for ListUsableWorkstations.

ListUsableWorkstationsResponse

Response message for ListUsableWorkstations.

ListWorkstationClustersRequest

Request message for ListWorkstationClusters.

ListWorkstationClustersResponse

Response message for ListWorkstationClusters.

ListWorkstationConfigsRequest

Request message for ListWorkstationConfigs.

ListWorkstationConfigsResponse

Response message for ListWorkstationConfigs.

ListWorkstationsRequest

Request message for ListWorkstations.

ListWorkstationsResponse

Response message for ListWorkstations.

OperationMetadata

Metadata for long-running operations.

StartWorkstationRequest

Request message for StartWorkstation.

StopWorkstationRequest

Request message for StopWorkstation.

UpdateWorkstationClusterRequest

Request message for UpdateWorkstationCluster.

UpdateWorkstationConfigRequest

Request message for UpdateWorkstationConfig.

UpdateWorkstationRequest

Request message for UpdateWorkstation.

Workstation

A single instance of a developer workstation with its own persistent storage.

WorkstationCluster

A workstation cluster resource in the Cloud Workstations API.

WorkstationConfig

A workstation configuration resource in the Cloud Workstations API.