Module google_api_proto::google::cloud::workstations::v1beta::workstation_config::persistent_directory

Nested message and enum types in PersistentDirectory.

Modules

gce_regional_persistent_disk

Nested message and enum types in GceRegionalPersistentDisk.

Structs

GceRegionalPersistentDisk

A PersistentDirectory backed by a Compute Engine regional persistent disk. The [persistent_directories][google.cloud.workstations.v1beta.WorkstationConfig.persistent_directories] field is repeated, but it may contain only one entry. It creates a persistent disk that mounts to the workstation VM at /home when the session starts and detaches when the session ends. If this field is empty, workstations created with this configuration do not have a persistent home directory.

Enums

DirectoryType

How a persistent directory should be implemented.