Struct google_api_proto::google::container::v1beta1::UpdateNodePoolRequest

pub struct UpdateNodePoolRequest {

    pub project_id: String,
    pub zone: String,
    pub cluster_id: String,
    pub node_pool_id: String,
    pub node_version: String,
    pub image_type: String,
    pub locations: Vec<String>,
    pub workload_metadata_config: Option<WorkloadMetadataConfig>,
    pub name: String,
    pub upgrade_settings: Option<UpgradeSettings>,
    pub tags: Option<NetworkTags>,
    pub taints: Option<NodeTaints>,
    pub labels: Option<NodeLabels>,
    pub linux_node_config: Option<LinuxNodeConfig>,
    pub kubelet_config: Option<NodeKubeletConfig>,
    pub node_network_config: Option<NodeNetworkConfig>,
    pub gcfs_config: Option<GcfsConfig>,
    pub confidential_nodes: Option<ConfidentialNodes>,
    pub gvnic: Option<VirtualNic>,
    pub etag: String,
    pub fast_socket: Option<FastSocket>,
    pub logging_config: Option<NodePoolLoggingConfig>,
    pub resource_labels: Option<ResourceLabels>,
    pub windows_node_config: Option<WindowsNodeConfig>,
    pub machine_type: String,
    pub disk_type: String,
    pub disk_size_gb: i64,
    pub resource_manager_tags: Option<ResourceManagerTags>,

}

SetNodePoolVersionRequest updates the version of a node pool.

Fields
project_id: String
๐Deprecated

Required. Deprecated. The Google Developers Console project ID or project number. This field has been deprecated and replaced by the name field.

zone: String
๐Deprecated

Required. Deprecated. The name of the Google Compute Engine zone in which the cluster resides. This field has been deprecated and replaced by the name field.

cluster_id: String
๐Deprecated

Required. Deprecated. The name of the cluster to upgrade. This field has been deprecated and replaced by the name field.

node_pool_id: String
๐Deprecated

Required. Deprecated. The name of the node pool to upgrade. This field has been deprecated and replaced by the name field.

node_version: String

Required. The Kubernetes version to change the nodes to (typically an upgrade).

Users may specify either explicit versions offered by Kubernetes Engine or version aliases, which have the following behavior:

โlatestโ: picks the highest valid Kubernetes version

โ1.Xโ: picks the highest valid patch+gke.N patch in the 1.X version

โ1.X.Yโ: picks the highest valid gke.N patch in the 1.X.Y version

โ1.X.Y-gke.Nโ: picks an explicit Kubernetes version

โ-โ: picks the Kubernetes master version

image_type: String

Required. The desired image type for the node pool. Please see https://cloud.google.com/kubernetes-engine/docs/concepts/node-images for available image types.

locations: Vec<String>

The desired list of Google Compute Engine zones in which the node poolโs nodes should be located. Changing the locations for a node pool will result in nodes being either created or removed from the node pool, depending on whether locations are being added or removed.

workload_metadata_config: Option<WorkloadMetadataConfig>

The desired workload metadata config for the node pool.

name: String

The name (project, location, cluster, node pool) of the node pool to update. Specified in the format projects/*/locations/*/clusters/*/nodePools/*.

upgrade_settings: Option<UpgradeSettings>

Upgrade settings control disruption and speed of the upgrade.

tags: Option<NetworkTags>

The desired network tags to be applied to all nodes in the node pool. If this field is not present, the tags will not be changed. Otherwise, the existing network tags will be replaced with the provided tags.

taints: Option<NodeTaints>

The desired node taints to be applied to all nodes in the node pool. If this field is not present, the taints will not be changed. Otherwise, the existing node taints will be replaced with the provided taints.

labels: Option<NodeLabels>

The desired node labels to be applied to all nodes in the node pool. If this field is not present, the labels will not be changed. Otherwise, the existing node labels will be replaced with the provided labels.

linux_node_config: Option<LinuxNodeConfig>

Parameters that can be configured on Linux nodes.

kubelet_config: Option<NodeKubeletConfig>

Node kubelet configs.

node_network_config: Option<NodeNetworkConfig>

Node network config.

gcfs_config: Option<GcfsConfig>

GCFS config.

confidential_nodes: Option<ConfidentialNodes>

Confidential nodes config. All the nodes in the node pool will be Confidential VM once enabled.

gvnic: Option<VirtualNic>

Enable or disable gvnic on the node pool.

etag: String

The current etag of the node pool. If an etag is provided and does not match the current etag of the node pool, update will be blocked and an ABORTED error will be returned.

fast_socket: Option<FastSocket>

Enable or disable NCCL fast socket for the node pool.

logging_config: Option<NodePoolLoggingConfig>

Logging configuration.

resource_labels: Option<ResourceLabels>

The resource labels for the node pool to use to annotate any related Google Compute Engine resources.

windows_node_config: Option<WindowsNodeConfig>

Parameters that can be configured on Windows nodes.

machine_type: String

Optional. The desired machine type for nodes in the node pool. Initiates an upgrade operation that migrates the nodes in the node pool to the specified machine type.

disk_type: String

Optional. The desired disk type for nodes in the node pool. Initiates an upgrade operation that migrates the nodes in the node pool to the specified disk type.

disk_size_gb: i64

Optional. The desired disk size for nodes in the node pool. Initiates an upgrade operation that migrates the nodes in the node pool to the specified disk size.

resource_manager_tags: Option<ResourceManagerTags>

Desired resource manager tag keys and values to be attached to the nodes for managing Compute Engine firewalls using Network Firewall Policies. Existing tags will be replaced with new values.