Module google_api_proto::google::devtools::artifactregistry::v1

Modules

apt_artifact

Nested message and enum types in AptArtifact.

artifact_registry_client

Generated client implementations.

cleanup_policy

Nested message and enum types in CleanupPolicy.

cleanup_policy_condition

Nested message and enum types in CleanupPolicyCondition.

hash

Nested message and enum types in Hash.

import_apt_artifacts_error_info

Nested message and enum types in ImportAptArtifactsErrorInfo.

import_apt_artifacts_request

Nested message and enum types in ImportAptArtifactsRequest.

import_yum_artifacts_error_info

Nested message and enum types in ImportYumArtifactsErrorInfo.

import_yum_artifacts_request

Nested message and enum types in ImportYumArtifactsRequest.

project_settings

Nested message and enum types in ProjectSettings.

remote_repository_config

Nested message and enum types in RemoteRepositoryConfig.

repository

Nested message and enum types in Repository.

vpcsc_config

Nested message and enum types in VPCSCConfig.

yum_artifact

Nested message and enum types in YumArtifact.

Structs

AptArtifact

A detailed representation of an Apt artifact. Information in the record is derived from the archive’s control file. See https://www.debian.org/doc/debian-policy/ch-controlfields.html

BatchDeleteVersionsMetadata

The metadata of an LRO from deleting multiple versions.

BatchDeleteVersionsRequest

The request to delete multiple versions across a repository.

CleanupPolicy

Artifact policy configuration for repository cleanup policies.

CleanupPolicyCondition

CleanupPolicyCondition is a set of conditions attached to a CleanupPolicy. If multiple entries are set, all must be satisfied for the condition to be satisfied.

CleanupPolicyMostRecentVersions

CleanupPolicyMostRecentVersions is an alternate condition of a CleanupPolicy for retaining a minimum number of versions.

CreateRepositoryRequest

The request to create a new repository.

CreateTagRequest

The request to create a new tag.

DeletePackageRequest

The request to delete a package.

DeleteRepositoryRequest

The request to delete a repository.

DeleteTagRequest

The request to delete a tag.

DeleteVersionRequest

The request to delete a version.

DockerImage

DockerImage represents a docker artifact. The following fields are returned as untyped metadata in the Version resource, using camelcase keys (i.e. metadata.imageSizeBytes):

File

Files store content that is potentially associated with Packages or Versions.

GetDockerImageRequest

The request to get docker images.

GetFileRequest

The request to retrieve a file.

GetMavenArtifactRequest

The request to get maven artifacts.

GetNpmPackageRequest

The request to get npm packages.

GetPackageRequest

The request to retrieve a package.

GetProjectSettingsRequest

Gets the redirection status for a project.

GetPythonPackageRequest

The request to get python packages.

GetRepositoryRequest

The request to retrieve a repository.

GetTagRequest

The request to retrieve a tag.

GetVersionRequest

The request to retrieve a version.

GetVpcscConfigRequest

Gets the VPC SC config for a project.

Hash

A hash of file content.

ImportAptArtifactsErrorInfo

Error information explaining why a package was not imported.

ImportAptArtifactsGcsSource

Google Cloud Storage location where the artifacts currently reside.

ImportAptArtifactsMetadata

The operation metadata for importing artifacts.

ImportAptArtifactsRequest

The request to import new apt artifacts.

ImportAptArtifactsResponse

The response message from importing APT artifacts.

ImportYumArtifactsErrorInfo

Error information explaining why a package was not imported.

ImportYumArtifactsGcsSource

Google Cloud Storage location where the artifacts currently reside.

ImportYumArtifactsMetadata

The operation metadata for importing artifacts.

ImportYumArtifactsRequest

The request to import new yum artifacts.

ImportYumArtifactsResponse

The response message from importing YUM artifacts.

ListDockerImagesRequest

The request to list docker images.

ListDockerImagesResponse

The response from listing docker images.

ListFilesRequest

The request to list files.

ListFilesResponse

The response from listing files.

ListMavenArtifactsRequest

The request to list maven artifacts.

ListMavenArtifactsResponse

The response from listing maven artifacts.

ListNpmPackagesRequest

The request to list npm packages.

ListNpmPackagesResponse

The response from listing npm packages.

ListPackagesRequest

The request to list packages.

ListPackagesResponse

The response from listing packages.

ListPythonPackagesRequest

The request to list python packages.

ListPythonPackagesResponse

The response from listing python packages.

ListRepositoriesRequest

The request to list repositories.

ListRepositoriesResponse

The response from listing repositories.

ListTagsRequest

The request to list tags.

ListTagsResponse

The response from listing tags.

ListVersionsRequest

The request to list versions.

ListVersionsResponse

The response from listing versions.

MavenArtifact

MavenArtifact represents a maven artifact.

NpmPackage

NpmPackage represents an npm artifact.

OperationMetadata

Metadata type for longrunning-operations, currently empty.

Package

Packages are named collections of versions.

ProjectSettings

The Artifact Registry settings that apply to a Project.

PythonPackage

PythonPackage represents a python artifact.

RemoteRepositoryConfig

Remote repository configuration.

Repository

A Repository for storing artifacts with a specific format.

Tag

Tags point to a version and represent an alternative name that can be used to access the version.

UpdateProjectSettingsRequest

Sets the settings of the project.

UpdateRepositoryRequest

The request to update a repository.

UpdateTagRequest

The request to create or update a tag.

UpdateVpcscConfigRequest

Sets the VPCSC config of the project.

UpstreamPolicy

Artifact policy configuration for the repository contents.

Version

The body of a version resource. A version resource represents a collection of components, such as files and other data. This may correspond to a version in many package management schemes.

VirtualRepositoryConfig

Virtual repository configuration.

VpcscConfig

The Artifact Registry VPC SC config that apply to a Project.

YumArtifact

A detailed representation of a Yum artifact.

Enums

VersionView

The view, which determines what version information is returned in a response.