Module google_api_proto::google::devtools::build::v1

Modules

build_event

Nested message and enum types in BuildEvent.

build_status

Nested message and enum types in BuildStatus.

publish_build_event_client

Generated client implementations.

publish_lifecycle_event_request

Nested message and enum types in PublishLifecycleEventRequest.

stream_id

Nested message and enum types in StreamId.

Structs

BuildEvent

An event representing some state change that occurred in the build. This message does not include field for uniquely identifying an event.

BuildStatus

Status used for both invocation attempt and overall build completion.

OrderedBuildEvent

Build event with contextual information about the stream it belongs to and its position in that stream.

PublishBuildToolEventStreamRequest

Streaming request message for PublishBuildToolEventStream.

PublishBuildToolEventStreamResponse

States which event has been committed. Any failure to commit will cause RPC errors, hence not recorded by this proto.

PublishLifecycleEventRequest

Publishes ‘lifecycle events’ that update the high-level state of a build:

StreamId

Unique identifier for a build event stream.

Enums

ConsoleOutputStream

The type of console output stream.