Module google_api_proto::google::devtools::cloudbuild::v2

Modules

connection

Nested message and enum types in Connection.

fetch_git_refs_request

Nested message and enum types in FetchGitRefsRequest.

installation_state

Nested message and enum types in InstallationState.

repository_manager_client

Generated client implementations.

Structs

BatchCreateRepositoriesRequest

Message for creating repositoritories in batch.

BatchCreateRepositoriesResponse

Message for response of creating repositories in batch.

Connection

A connection to a SCM like GitHub, GitHub Enterprise, Bitbucket Server or GitLab.

CreateConnectionRequest

Message for creating a Connection

CreateRepositoryRequest

Message for creating a Repository.

DeleteConnectionRequest

Message for deleting a Connection.

DeleteRepositoryRequest

Message for deleting a Repository.

FetchGitRefsRequest

Request for fetching git refs

FetchGitRefsResponse

Response for fetching git refs

FetchLinkableRepositoriesRequest

Request message for FetchLinkableRepositories.

FetchLinkableRepositoriesResponse

Response message for FetchLinkableRepositories.

FetchReadTokenRequest

Message for fetching SCM read token.

FetchReadTokenResponse

Message for responding to get read token.

FetchReadWriteTokenRequest

Message for fetching SCM read/write token.

FetchReadWriteTokenResponse

Message for responding to get read/write token.

GetConnectionRequest

Message for getting the details of a Connection.

GetRepositoryRequest

Message for getting the details of a Repository.

GitHubConfig

Configuration for connections to github.com.

GitHubEnterpriseConfig

Configuration for connections to an instance of GitHub Enterprise.

GitLabConfig

Configuration for connections to gitlab.com or an instance of GitLab Enterprise.

InstallationState

Describes stage and necessary actions to be taken by the user to complete the installation. Used for GitHub and GitHub Enterprise based connections.

ListConnectionsRequest

Message for requesting list of Connections.

ListConnectionsResponse

Message for response to listing Connections.

ListRepositoriesRequest

Message for requesting list of Repositories.

ListRepositoriesResponse

Message for response to listing Repositories.

OAuthCredential

Represents an OAuth token of the account that authorized the Connection, and associated metadata.

OperationMetadata

Represents the metadata of the long-running operation.

ProcessWebhookRequest

RPC request object accepted by the ProcessWebhook RPC method.

Repository

A repository associated to a parent connection.

RunWorkflowCustomOperationMetadata

Represents the custom metadata of the RunWorkflow long-running operation.

ServiceDirectoryConfig

ServiceDirectoryConfig represents Service Directory configuration for a connection.

UpdateConnectionRequest

Message for updating a Connection.

UserCredential

Represents a personal access token that authorized the Connection, and associated metadata.