Module google_api_proto::google::devtools::clouddebugger::v2

Modules

breakpoint

Nested message and enum types in Breakpoint.

controller2_client

Generated client implementations.

debugger2_client

Generated client implementations.

list_breakpoints_request

Nested message and enum types in ListBreakpointsRequest.

status_message

Nested message and enum types in StatusMessage.

Structs

Breakpoint

Represents the breakpoint specification, status and results.

Debuggee

Represents the debugged application. The application may include one or more replicated processes executing the same code. Each of these processes is attached with a debugger agent, carrying out the debugging commands. Agents attached to the same debuggee identify themselves as such by using exactly the same Debuggee message value when registering.

DeleteBreakpointRequest

Request to delete a breakpoint.

FormatMessage

Represents a message with parameters.

GetBreakpointRequest

Request to get breakpoint information.

GetBreakpointResponse

Response for getting breakpoint information.

ListActiveBreakpointsRequest

Request to list active breakpoints.

ListActiveBreakpointsResponse

Response for listing active breakpoints.

ListBreakpointsRequest

Request to list breakpoints.

ListBreakpointsResponse

Response for listing breakpoints.

ListDebuggeesRequest

Request to list debuggees.

ListDebuggeesResponse

Response for listing debuggees.

RegisterDebuggeeRequest

Request to register a debuggee.

RegisterDebuggeeResponse

Response for registering a debuggee.

SetBreakpointRequest

Request to set a breakpoint

SetBreakpointResponse

Response for setting a breakpoint.

SourceLocation

Represents a location in the source code.

StackFrame

Represents a stack frame context.

StatusMessage

Represents a contextual status message. The message can indicate an error or informational status, and refer to specific parts of the containing object. For example, the Breakpoint.status field can indicate an error referring to the BREAKPOINT_SOURCE_LOCATION with the message Location not found.

UpdateActiveBreakpointRequest

Request to update an active breakpoint.

UpdateActiveBreakpointResponse

Response for updating an active breakpoint. The message is defined to allow future extensions.

Variable

Represents a variable or an argument possibly of a compound object type. Note how the following variables are represented: