Module google_api_proto::google::devtools::cloudtrace::v2

Modules

attribute_value

Nested message and enum types in AttributeValue.

span

Nested message and enum types in Span.

stack_trace

Nested message and enum types in StackTrace.

trace_service_client

Generated client implementations.

Structs

AttributeValue

The allowed types for [VALUE] in a \[KEY]:[VALUE\] attribute.

BatchWriteSpansRequest

The request message for the BatchWriteSpans method.

Module

Binary module.

Span

A span represents a single operation within a trace. Spans can be nested to form a trace tree. Often, a trace contains a root span that describes the end-to-end latency, and one or more subspans for its sub-operations. A trace can also contain multiple root spans, or none at all. Spans do not need to be contiguous—there may be gaps or overlaps between spans in a trace.

StackTrace

A call stack appearing in a trace.

TruncatableString

Represents a string that might be shortened to a specified length.