Module google_api_proto::google::home::graph::v1

Modules

home_graph_api_service_client

Generated client implementations.

Structs

AgentDeviceId

Third-party device ID for one device.

AgentOtherDeviceId

Alternate third-party device ID.

DeleteAgentUserRequest

Request type for the DeleteAgentUser call.

Device

Third-party device definition.

DeviceInfo

Device information.

DeviceNames

Identifiers used to describe the device.

QueryRequest

Request type for the Query call.

QueryRequestInput

Device ID inputs to [QueryRequest][google.home.graph.v1.QueryRequest].

QueryRequestPayload

Payload containing device IDs.

QueryResponse

Response type for the Query call. This should follow the same format as the Google smart home action.devices.QUERY response.

QueryResponsePayload

Payload containing device states information.

ReportStateAndNotificationDevice

The states and notifications specific to a device.

ReportStateAndNotificationRequest

Request type for the ReportStateAndNotification call. It may include states, notifications, or both. States and notifications are defined per device_id (for example, “123” and “456” in the following example).

ReportStateAndNotificationResponse

Response type for the ReportStateAndNotification call.

RequestSyncDevicesRequest

Request type for the RequestSyncDevices call.

RequestSyncDevicesResponse

Response type for the RequestSyncDevices call.

StateAndNotificationPayload

Payload containing the state and notification information for devices.

SyncRequest

Request type for the Sync call.

SyncResponse

Response type for the Sync call. This should follow the same format as the Google smart home action.devices.SYNC response.

SyncResponsePayload

Payload containing device information.