Module google_api_proto::google::iam::v1

Modules

audit_config_delta

Nested message and enum types in AuditConfigDelta.

audit_log_config

Nested message and enum types in AuditLogConfig.

binding_delta

Nested message and enum types in BindingDelta.

iam_policy_client

Generated client implementations.

logging

Structs

AuditConfig

Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs.

AuditConfigDelta

One delta entry for AuditConfig. Each individual change (only one exempted_member in each entry) to a AuditConfig will be a separate entry.

AuditLogConfig

Provides the configuration for logging a type of permissions. Example:

Binding

Associates members, or principals, with a role.

BindingDelta

One delta entry for Binding. Each individual change (only one member in each entry) to a binding will be a separate entry.

GetIamPolicyRequest

Request message for GetIamPolicy method.

GetPolicyOptions

Encapsulates settings provided to GetIamPolicy.

Policy

An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources.

PolicyDelta

The difference delta between two policies.

SetIamPolicyRequest

Request message for SetIamPolicy method.

TestIamPermissionsRequest

Request message for TestIamPermissions method.

TestIamPermissionsResponse

Response message for TestIamPermissions method.