Struct google_api_proto::google::iam::v1beta::WorkloadIdentityPoolProvider

pub struct WorkloadIdentityPoolProvider {
    pub name: String,
    pub display_name: String,
    pub description: String,
    pub state: i32,
    pub disabled: bool,
    pub attribute_mapping: BTreeMap<String, String>,
    pub attribute_condition: String,
    pub provider_config: Option<ProviderConfig>,
}

A configuration for an external identity provider.

Fields
name: String

Output only. The resource name of the provider.

display_name: String

A display name for the provider. Cannot exceed 32 characters.

description: String

A description for the provider. Cannot exceed 256 characters.

state: i32

Output only. The state of the provider.

disabled: bool

Whether the provider is disabled. You cannot use a disabled provider to exchange tokens. However, existing tokens still grant access.

attribute_mapping: BTreeMap<String, String>

Maps attributes from authentication credentials issued by an external identity provider to Google Cloud attributes, such as subject and segment.

Each key must be a string specifying the Google Cloud IAM attribute to map to.

The following keys are supported:

google.subject: The principal IAM is authenticating. You can reference this value in IAM bindings. This is also the subject that appears in Cloud Logging logs. Cannot exceed 127 characters.

google.groups: Groups the external identity belongs to. You can grant groups access to resources using an IAM principalSet binding; access applies to all members of the group.

You can also provide custom attributes by specifying attribute.{custom_attribute}, where {custom_attribute} is the name of the custom attribute to be mapped. You can define a maximum of 50 custom attributes. The maximum length of a mapped attribute key is 100 characters, and the key may only contain the characters [a-z0-9_].

You can reference these attributes in IAM policies to define fine-grained access for a workload to Google Cloud resources. For example:

google.subject: principal://iam.googleapis.com/projects/{project}/locations/{location}/workloadIdentityPools/{pool}/subject/{value}

google.groups: principalSet://iam.googleapis.com/projects/{project}/locations/{location}/workloadIdentityPools/{pool}/group/{value}

attribute.{custom_attribute}: principalSet://iam.googleapis.com/projects/{project}/locations/{location}/workloadIdentityPools/{pool}/attribute.{custom_attribute}/{value}

Each value must be a [Common Expression Language] (https://opensource.google/projects/cel) function that maps an identity provider credential to the normalized attribute specified by the corresponding map key.

You can use the assertion keyword in the expression to access a JSON representation of the authentication credential issued by the provider.

The maximum length of an attribute mapping expression is 2048 characters. When evaluated, the total size of all mapped attributes must not exceed 8KB.

For AWS providers, the following rules apply:

If no attribute mapping is defined, the following default mapping applies:

{
  "google.subject":"assertion.arn",
  "attribute.aws_role":
      "assertion.arn.contains('assumed-role')"
      " ? assertion.arn.extract('{account_arn}assumed-role/')"
      "   + 'assumed-role/'"
      "   + assertion.arn.extract('assumed-role/{role_name}/')"
      " : assertion.arn",
}

If any custom attribute mappings are defined, they must include a mapping to the google.subject attribute.

For OIDC providers, the following rules apply:

Custom attribute mappings must be defined, and must include a mapping to the google.subject attribute. For example, the following maps the sub claim of the incoming credential to the subject attribute on a Google token.

{"google.subject": "assertion.sub"}

attribute_condition: String

A Common Expression Language expression, in plain text, to restrict what otherwise valid authentication credentials issued by the provider should not be accepted.

The expression must output a boolean representing whether to allow the federation.

The following keywords may be referenced in the expressions:

assertion: JSON representing the authentication credential issued by the provider.

google: The Google attributes mapped from the assertion in the attribute_mappings.

attribute: The custom attributes mapped from the assertion in the attribute_mappings.

The maximum length of the attribute condition expression is 4096 characters. If unspecified, all valid authentication credential are accepted.

The following example shows how to only allow credentials with a mapped google.groups value of admins:

"'admins' in google.groups"

provider_config: Option<ProviderConfig>

Identity provider configuration types.

Implementations

impl WorkloadIdentityPoolProvider

pub fn state(&self) -> State

Returns the enum value of state, or the default if the field is set to an invalid enum value.

pub fn set_state(&mut self, value: State)

Sets state to the provided enum value.