Struct google_api_proto::google::identity::accesscontextmanager::v1::access_context_manager_client::AccessContextManagerClient

pub struct AccessContextManagerClient<T> { /* private fields */ }

API for setting [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] and [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] for Google Cloud projects. Each organization has one [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] that contains the [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] and [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter]. This [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] is applicable to all resources in the organization. AccessPolicies

Implementations

impl<T> AccessContextManagerClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> AccessContextManagerClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn list_access_policies( &mut self, request: impl IntoRequest<ListAccessPoliciesRequest> ) -> Result<Response<ListAccessPoliciesResponse>, Status>

Lists all [access policies] [google.identity.accesscontextmanager.v1.AccessPolicy] in an organization.

pub async fn get_access_policy( &mut self, request: impl IntoRequest<GetAccessPolicyRequest> ) -> Result<Response<AccessPolicy>, Status>

Returns an [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] based on the name.

pub async fn create_access_policy( &mut self, request: impl IntoRequest<AccessPolicy> ) -> Result<Response<Operation>, Status>

Creates an access policy. This method fails if the organization already has an access policy. The long-running operation has a successful status after the access policy propagates to long-lasting storage. Syntactic and basic semantic errors are returned in metadata as a BadRequest proto.

pub async fn update_access_policy( &mut self, request: impl IntoRequest<UpdateAccessPolicyRequest> ) -> Result<Response<Operation>, Status>

Updates an [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy]. The long-running operation from this RPC has a successful status after the changes to the [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] propagate to long-lasting storage.

pub async fn delete_access_policy( &mut self, request: impl IntoRequest<DeleteAccessPolicyRequest> ) -> Result<Response<Operation>, Status>

Deletes an [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] based on the resource name. The long-running operation has a successful status after the [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] is removed from long-lasting storage.

pub async fn list_access_levels( &mut self, request: impl IntoRequest<ListAccessLevelsRequest> ) -> Result<Response<ListAccessLevelsResponse>, Status>

Lists all [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] for an access policy.

pub async fn get_access_level( &mut self, request: impl IntoRequest<GetAccessLevelRequest> ) -> Result<Response<AccessLevel>, Status>

Gets an [access level] [google.identity.accesscontextmanager.v1.AccessLevel] based on the resource name.

pub async fn create_access_level( &mut self, request: impl IntoRequest<CreateAccessLevelRequest> ) -> Result<Response<Operation>, Status>

Creates an [access level] [google.identity.accesscontextmanager.v1.AccessLevel]. The long-running operation from this RPC has a successful status after the [access level] [google.identity.accesscontextmanager.v1.AccessLevel] propagates to long-lasting storage. If [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] contain errors, an error response is returned for the first error encountered.

pub async fn update_access_level( &mut self, request: impl IntoRequest<UpdateAccessLevelRequest> ) -> Result<Response<Operation>, Status>

Updates an [access level] [google.identity.accesscontextmanager.v1.AccessLevel]. The long-running operation from this RPC has a successful status after the changes to the [access level] [google.identity.accesscontextmanager.v1.AccessLevel] propagate to long-lasting storage. If [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] contain errors, an error response is returned for the first error encountered.

pub async fn delete_access_level( &mut self, request: impl IntoRequest<DeleteAccessLevelRequest> ) -> Result<Response<Operation>, Status>

Deletes an [access level] [google.identity.accesscontextmanager.v1.AccessLevel] based on the resource name. The long-running operation from this RPC has a successful status after the [access level] [google.identity.accesscontextmanager.v1.AccessLevel] has been removed from long-lasting storage.

pub async fn replace_access_levels( &mut self, request: impl IntoRequest<ReplaceAccessLevelsRequest> ) -> Result<Response<Operation>, Status>

Replaces all existing [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] in an [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] with the [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] provided. This is done atomically. The long-running operation from this RPC has a successful status after all replacements propagate to long-lasting storage. If the replacement contains errors, an error response is returned for the first error encountered. Upon error, the replacement is cancelled, and existing [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] are not affected. The Operation.response field contains ReplaceAccessLevelsResponse. Removing [access levels] [google.identity.accesscontextmanager.v1.AccessLevel] contained in existing [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] result in an error.

pub async fn list_service_perimeters( &mut self, request: impl IntoRequest<ListServicePerimetersRequest> ) -> Result<Response<ListServicePerimetersResponse>, Status>

Lists all [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] for an access policy.

pub async fn get_service_perimeter( &mut self, request: impl IntoRequest<GetServicePerimeterRequest> ) -> Result<Response<ServicePerimeter>, Status>

Gets a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] based on the resource name.

pub async fn create_service_perimeter( &mut self, request: impl IntoRequest<CreateServicePerimeterRequest> ) -> Result<Response<Operation>, Status>

Creates a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter]. The long-running operation from this RPC has a successful status after the [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] propagates to long-lasting storage. If a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] contains errors, an error response is returned for the first error encountered.

pub async fn update_service_perimeter( &mut self, request: impl IntoRequest<UpdateServicePerimeterRequest> ) -> Result<Response<Operation>, Status>

Updates a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter]. The long-running operation from this RPC has a successful status after the [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] propagates to long-lasting storage. If a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] contains errors, an error response is returned for the first error encountered.

pub async fn delete_service_perimeter( &mut self, request: impl IntoRequest<DeleteServicePerimeterRequest> ) -> Result<Response<Operation>, Status>

Deletes a [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] based on the resource name. The long-running operation from this RPC has a successful status after the [service perimeter] [google.identity.accesscontextmanager.v1.ServicePerimeter] is removed from long-lasting storage.

pub async fn replace_service_perimeters( &mut self, request: impl IntoRequest<ReplaceServicePerimetersRequest> ) -> Result<Response<Operation>, Status>

Replace all existing [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] in an [access policy] [google.identity.accesscontextmanager.v1.AccessPolicy] with the [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] provided. This is done atomically. The long-running operation from this RPC has a successful status after all replacements propagate to long-lasting storage. Replacements containing errors result in an error response for the first error encountered. Upon an error, replacement are cancelled and existing [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] are not affected. The Operation.response field contains ReplaceServicePerimetersResponse.

pub async fn commit_service_perimeters( &mut self, request: impl IntoRequest<CommitServicePerimetersRequest> ) -> Result<Response<Operation>, Status>

Commits the dry-run specification for all the [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] in an [access policy][google.identity.accesscontextmanager.v1.AccessPolicy]. A commit operation on a service perimeter involves copying its spec field to the status field of the service perimeter. Only [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] with use_explicit_dry_run_spec field set to true are affected by a commit operation. The long-running operation from this RPC has a successful status after the dry-run specifications for all the [service perimeters] [google.identity.accesscontextmanager.v1.ServicePerimeter] have been committed. If a commit fails, it causes the long-running operation to return an error response and the entire commit operation is cancelled. When successful, the Operation.response field contains CommitServicePerimetersResponse. The dry_run and the spec fields are cleared after a successful commit operation.

pub async fn list_gcp_user_access_bindings( &mut self, request: impl IntoRequest<ListGcpUserAccessBindingsRequest> ) -> Result<Response<ListGcpUserAccessBindingsResponse>, Status>

Lists all [GcpUserAccessBindings] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding] for a Google Cloud organization.

pub async fn get_gcp_user_access_binding( &mut self, request: impl IntoRequest<GetGcpUserAccessBindingRequest> ) -> Result<Response<GcpUserAccessBinding>, Status>

Gets the [GcpUserAccessBinding] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding] with the given name.

pub async fn create_gcp_user_access_binding( &mut self, request: impl IntoRequest<CreateGcpUserAccessBindingRequest> ) -> Result<Response<Operation>, Status>

Creates a [GcpUserAccessBinding] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding]. If the client specifies a [name] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding.name], the server ignores it. Fails if a resource already exists with the same [group_key] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding.group_key]. Completion of this long-running operation does not necessarily signify that the new binding is deployed onto all affected users, which may take more time.

pub async fn update_gcp_user_access_binding( &mut self, request: impl IntoRequest<UpdateGcpUserAccessBindingRequest> ) -> Result<Response<Operation>, Status>

Updates a [GcpUserAccessBinding] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding]. Completion of this long-running operation does not necessarily signify that the changed binding is deployed onto all affected users, which may take more time.

pub async fn delete_gcp_user_access_binding( &mut self, request: impl IntoRequest<DeleteGcpUserAccessBindingRequest> ) -> Result<Response<Operation>, Status>

Deletes a [GcpUserAccessBinding] [google.identity.accesscontextmanager.v1.GcpUserAccessBinding]. Completion of this long-running operation does not necessarily signify that the binding deletion is deployed onto all affected users, which may take more time.

pub async fn set_iam_policy( &mut self, request: impl IntoRequest<SetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Sets the IAM policy for the specified Access Context Manager [access policy][google.identity.accesscontextmanager.v1.AccessPolicy]. This method replaces the existing IAM policy on the access policy. The IAM policy controls the set of users who can perform specific operations on the Access Context Manager [access policy][google.identity.accesscontextmanager.v1.AccessPolicy].

pub async fn get_iam_policy( &mut self, request: impl IntoRequest<GetIamPolicyRequest> ) -> Result<Response<Policy>, Status>

Gets the IAM policy for the specified Access Context Manager [access policy][google.identity.accesscontextmanager.v1.AccessPolicy].

pub async fn test_iam_permissions( &mut self, request: impl IntoRequest<TestIamPermissionsRequest> ) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the IAM permissions that the caller has on the specified Access Context Manager resource. The resource can be an [AccessPolicy][google.identity.accesscontextmanager.v1.AccessPolicy], [AccessLevel][google.identity.accesscontextmanager.v1.AccessLevel], or [ServicePerimeter][google.identity.accesscontextmanager.v1.ServicePerimeter ]. This method does not support other resources.