Struct google_api_proto::google::maps::places::v1::Place

pub struct Place {

    pub id: String,
    pub display_name: Option<LocalizedText>,
    pub types: Vec<String>,
    pub national_phone_number: String,
    pub international_phone_number: String,
    pub formatted_address: String,
    pub address_components: Vec<AddressComponent>,
    pub plus_code: Option<PlusCode>,
    pub location: Option<LatLng>,
    pub viewport: Option<Viewport>,
    pub rating: f64,
    pub google_maps_uri: String,
    pub website_uri: String,
    pub reviews: Vec<Review>,
    pub opening_hours: Option<OpeningHours>,
    pub utc_offset_minutes: i32,
    pub adr_format_address: String,
    pub business_status: i32,
    pub price_level: i32,
    pub attributions: Vec<Attribution>,
    pub user_rating_count: i32,
    pub icon_mask_base_uri: String,
    pub icon_background_color: String,
    pub takeout: Option<bool>,
    pub delivery: Option<bool>,
    pub dine_in: Option<bool>,
    pub curbside_pickup: Option<bool>,
    pub wheelchair_accessible_entrance: Option<bool>,
    pub reservable: Option<bool>,
    pub serves_breakfast: Option<bool>,
    pub serves_lunch: Option<bool>,
    pub serves_dinner: Option<bool>,
    pub serves_beer: Option<bool>,
    pub serves_wine: Option<bool>,
    pub serves_brunch: Option<bool>,
    pub serves_vegetarian_food: Option<bool>,
    pub current_opening_hours: Option<OpeningHours>,
    pub current_secondary_opening_hours: Vec<OpeningHours>,
    pub secondary_opening_hours: Vec<OpeningHours>,
    pub editorial_summary: Option<EditorialSummary>,

}

All the information representing a Place.

Fields
id: String

Required. The unique identifier of a place.

display_name: Option<LocalizedText>

The localized name of the place, suitable as a short human-readable description. For example, “Google Sydney”, “Starbucks”, “Pyrmont”, etc.

types: Vec<String>

A set of type tags for this result. For example, “political” and “locality”.

national_phone_number: String

A human-readable phone number for the place, in national format.

international_phone_number: String

A human-readable phone number for the place, in international format.

formatted_address: String

A full, human-readable address for this place.

address_components: Vec<AddressComponent>

Repeated components for each locality level.

plus_code: Option<PlusCode>

Plus code of the place location lat/long.

location: Option<LatLng>

The position of this place.

viewport: Option<Viewport>

A viewport suitable for displaying the place on an average-sized map.

rating: f64

A rating between 1.0 and 5.0, based on user reviews of this place.

google_maps_uri: String

A URL providing more information about this place.

website_uri: String

The authoritative website for this place, e.g. a business’ homepage. Note that for places that are part of a chain (e.g. an IKEA store), this will usually be the website for the individual store, not the overall chain.

reviews: Vec<Review>

List of reviews about this place.

opening_hours: Option<OpeningHours>

The regular hours of operation.

utc_offset_minutes: i32

Number of minutes this place’s timezone is currently offset from UTC. This is expressed in minutes to support timezones that are offset by fractions of an hour, e.g. X hours and 15 minutes.

adr_format_address: String

The place’s address in adr microformat: http://microformats.org/wiki/adr.

business_status: i32

The business status for the place.

price_level: i32

Price level of the place.

attributions: Vec<Attribution>

A set of data provider that must be shown with this result.

user_rating_count: i32

The total number of reviews (with or without text) for this place.

icon_mask_base_uri: String

A truncated URL to an v2 icon mask. User can access different icon type by appending type suffix to the end (eg, “.svg” or “.png”).

icon_background_color: String

Background color for icon_mask in hex format, e.g. #909CE1.

takeout: Option<bool>

Specifies if the business supports takeout.

delivery: Option<bool>

Specifies if the business supports delivery.

dine_in: Option<bool>

Specifies if the business supports indoor or outdoor seating options.

curbside_pickup: Option<bool>

Specifies if the business supports curbside pickup.

wheelchair_accessible_entrance: Option<bool>

Specifies if the place has an entrance that is wheelchair-accessible.

reservable: Option<bool>

Specifies if the place supports reservations.

serves_breakfast: Option<bool>

Specifies if the place serves breakfast.

serves_lunch: Option<bool>

Specifies if the place serves lunch.

serves_dinner: Option<bool>

Specifies if the place serves dinner.

serves_beer: Option<bool>

Specifies if the place serves beer.

serves_wine: Option<bool>

Specifies if the place serves wine.

serves_brunch: Option<bool>

Specifies if the place serves brunch.

serves_vegetarian_food: Option<bool>

Specifies if the place serves vegetarian food.

current_opening_hours: Option<OpeningHours>

The hours of operation for the next seven days (including today). The time period starts at midnight on the date of the request and ends at 11:59 pm six days later. This field includes the special_days subfield of all hours, set for dates that have exceptional hours.

current_secondary_opening_hours: Vec<OpeningHours>

Contains an array of entries for the next seven days including information about secondary hours of a business. Secondary hours are different from a business’s main hours. For example, a restaurant can specify drive through hours or delivery hours as its secondary hours. This field populates the type subfield, which draws from a predefined list of opening hours types (such as DRIVE_THROUGH, PICKUP, or TAKEOUT) based on the types of the place. This field includes the special_days subfield of all hours, set for dates that have exceptional hours.

secondary_opening_hours: Vec<OpeningHours>

Contains an array of entries for information about regular secondary hours of a business. Secondary hours are different from a business’s main hours. For example, a restaurant can specify drive through hours or delivery hours as its secondary hours. This field populates the type subfield, which draws from a predefined list of opening hours types (such as DRIVE_THROUGH, PICKUP, or TAKEOUT) based on the types of the place.

editorial_summary: Option<EditorialSummary>

Contains a summary of the place. A summary is comprised of a textual overview, and also includes the language code for these if applicable. Summary text must be presented as-is and can not be modified or altered.

Implementations

impl Place

pub fn business_status(&self) -> BusinessStatus

Returns the enum value of business_status, or the default if the field is set to an invalid enum value.

pub fn set_business_status(&mut self, value: BusinessStatus)

Sets business_status to the provided enum value.

pub fn price_level(&self) -> PriceLevel

Returns the enum value of price_level, or the default if the field is set to an invalid enum value.

pub fn set_price_level(&mut self, value: PriceLevel)

Sets price_level to the provided enum value.

pub fn takeout(&self) -> bool

Returns the value of takeout, or the default value if takeout is unset.

pub fn delivery(&self) -> bool

Returns the value of delivery, or the default value if delivery is unset.

pub fn dine_in(&self) -> bool

Returns the value of dine_in, or the default value if dine_in is unset.

pub fn curbside_pickup(&self) -> bool

Returns the value of curbside_pickup, or the default value if curbside_pickup is unset.

pub fn wheelchair_accessible_entrance(&self) -> bool

Returns the value of wheelchair_accessible_entrance, or the default value if wheelchair_accessible_entrance is unset.

pub fn reservable(&self) -> bool

Returns the value of reservable, or the default value if reservable is unset.

pub fn serves_breakfast(&self) -> bool

Returns the value of serves_breakfast, or the default value if serves_breakfast is unset.

pub fn serves_lunch(&self) -> bool

Returns the value of serves_lunch, or the default value if serves_lunch is unset.

pub fn serves_dinner(&self) -> bool

Returns the value of serves_dinner, or the default value if serves_dinner is unset.

pub fn serves_beer(&self) -> bool

Returns the value of serves_beer, or the default value if serves_beer is unset.

pub fn serves_wine(&self) -> bool

Returns the value of serves_wine, or the default value if serves_wine is unset.

pub fn serves_brunch(&self) -> bool

Returns the value of serves_brunch, or the default value if serves_brunch is unset.

pub fn serves_vegetarian_food(&self) -> bool

Returns the value of serves_vegetarian_food, or the default value if serves_vegetarian_food is unset.