Struct google_api_proto::google::maps::places::v1::Place

pub struct Place {

    pub name: String,
    pub id: String,
    pub display_name: Option<LocalizedText>,
    pub types: Vec<String>,
    pub national_phone_number: String,
    pub international_phone_number: String,
    pub formatted_address: String,
    pub address_components: Vec<AddressComponent>,
    pub plus_code: Option<PlusCode>,
    pub location: Option<LatLng>,
    pub viewport: Option<Viewport>,
    pub rating: f64,
    pub google_maps_uri: String,
    pub website_uri: String,
    pub reviews: Vec<Review>,
    pub regular_opening_hours: Option<OpeningHours>,
    pub utc_offset_minutes: Option<i32>,
    pub photos: Vec<Photo>,
    pub adr_format_address: String,
    pub business_status: i32,
    pub price_level: i32,
    pub attributions: Vec<Attribution>,
    pub user_rating_count: Option<i32>,
    pub icon_mask_base_uri: String,
    pub icon_background_color: String,
    pub takeout: Option<bool>,
    pub delivery: Option<bool>,
    pub dine_in: Option<bool>,
    pub curbside_pickup: Option<bool>,
    pub reservable: Option<bool>,
    pub serves_breakfast: Option<bool>,
    pub serves_lunch: Option<bool>,
    pub serves_dinner: Option<bool>,
    pub serves_beer: Option<bool>,
    pub serves_wine: Option<bool>,
    pub serves_brunch: Option<bool>,
    pub serves_vegetarian_food: Option<bool>,
    pub current_opening_hours: Option<OpeningHours>,
    pub current_secondary_opening_hours: Vec<OpeningHours>,
    pub regular_secondary_opening_hours: Vec<OpeningHours>,
    pub editorial_summary: Option<LocalizedText>,
    pub outdoor_seating: Option<bool>,
    pub live_music: Option<bool>,
    pub menu_for_children: Option<bool>,
    pub serves_cocktails: Option<bool>,
    pub serves_dessert: Option<bool>,
    pub serves_coffee: Option<bool>,
    pub good_for_children: Option<bool>,
    pub allows_dogs: Option<bool>,
    pub restroom: Option<bool>,
    pub good_for_groups: Option<bool>,
    pub good_for_watching_sports: Option<bool>,
    pub payment_options: Option<PaymentOptions>,
    pub parking_options: Option<ParkingOptions>,
    pub sub_destinations: Vec<SubDestination>,
    pub accessibility_options: Option<AccessibilityOptions>,
    pub fuel_options: Option<FuelOptions>,
    pub ev_charge_options: Option<EvChargeOptions>,

}

All the information representing a Place.

Fields
name: String

Output only. An ID representing this place which may be used to look up this place again (a.k.a. the API “resource” name: places/<place_id>).

id: String

Output only. The unique identifier of a place.

display_name: Option<LocalizedText>

Output only. The localized name of the place, suitable as a short human-readable description. For example, “Google Sydney”, “Starbucks”, “Pyrmont”, etc.

types: Vec<String>

Output only. A set of type tags for this result. For example, “political” and “locality”.

national_phone_number: String

Output only. A human-readable phone number for the place, in national format.

international_phone_number: String

Output only. A human-readable phone number for the place, in international format.

formatted_address: String

Output only. A full, human-readable address for this place.

address_components: Vec<AddressComponent>

Output only. Repeated components for each locality level.

plus_code: Option<PlusCode>

Output only. Plus code of the place location lat/long.

location: Option<LatLng>

Output only. The position of this place.

viewport: Option<Viewport>

Output only. A viewport suitable for displaying the place on an average-sized map.

rating: f64

Output only. A rating between 1.0 and 5.0, based on user reviews of this place.

google_maps_uri: String

Output only. A URL providing more information about this place.

website_uri: String

Output only. The authoritative website for this place, e.g. a business’ homepage. Note that for places that are part of a chain (e.g. an IKEA store), this will usually be the website for the individual store, not the overall chain.

reviews: Vec<Review>

Output only. List of reviews about this place.

regular_opening_hours: Option<OpeningHours>

Output only. The regular hours of operation.

utc_offset_minutes: Option<i32>

Output only. Number of minutes this place’s timezone is currently offset from UTC. This is expressed in minutes to support timezones that are offset by fractions of an hour, e.g. X hours and 15 minutes.

photos: Vec<Photo>

Output only. Information (including references) about photos of this place.

adr_format_address: String

Output only. The place’s address in adr microformat: http://microformats.org/wiki/adr.

business_status: i32

Output only. The business status for the place.

price_level: i32

Output only. Price level of the place.

attributions: Vec<Attribution>

Output only. A set of data provider that must be shown with this result.

user_rating_count: Option<i32>

Output only. The total number of reviews (with or without text) for this place.

icon_mask_base_uri: String

Output only. A truncated URL to an v2 icon mask. User can access different icon type by appending type suffix to the end (eg, “.svg” or “.png”).

icon_background_color: String

Output only. Background color for icon_mask in hex format, e.g. #909CE1.

takeout: Option<bool>

Output only. Specifies if the business supports takeout.

delivery: Option<bool>

Output only. Specifies if the business supports delivery.

dine_in: Option<bool>

Output only. Specifies if the business supports indoor or outdoor seating options.

curbside_pickup: Option<bool>

Output only. Specifies if the business supports curbside pickup.

reservable: Option<bool>

Output only. Specifies if the place supports reservations.

serves_breakfast: Option<bool>

Output only. Specifies if the place serves breakfast.

serves_lunch: Option<bool>

Output only. Specifies if the place serves lunch.

serves_dinner: Option<bool>

Output only. Specifies if the place serves dinner.

serves_beer: Option<bool>

Output only. Specifies if the place serves beer.

serves_wine: Option<bool>

Output only. Specifies if the place serves wine.

serves_brunch: Option<bool>

Output only. Specifies if the place serves brunch.

serves_vegetarian_food: Option<bool>

Output only. Specifies if the place serves vegetarian food.

current_opening_hours: Option<OpeningHours>

Output only. The hours of operation for the next seven days (including today). The time period starts at midnight on the date of the request and ends at 11:59 pm six days later. This field includes the special_days subfield of all hours, set for dates that have exceptional hours.

current_secondary_opening_hours: Vec<OpeningHours>

Output only. Contains an array of entries for the next seven days including information about secondary hours of a business. Secondary hours are different from a business’s main hours. For example, a restaurant can specify drive through hours or delivery hours as its secondary hours. This field populates the type subfield, which draws from a predefined list of opening hours types (such as DRIVE_THROUGH, PICKUP, or TAKEOUT) based on the types of the place. This field includes the special_days subfield of all hours, set for dates that have exceptional hours.

regular_secondary_opening_hours: Vec<OpeningHours>

Output only. Contains an array of entries for information about regular secondary hours of a business. Secondary hours are different from a business’s main hours. For example, a restaurant can specify drive through hours or delivery hours as its secondary hours. This field populates the type subfield, which draws from a predefined list of opening hours types (such as DRIVE_THROUGH, PICKUP, or TAKEOUT) based on the types of the place.

editorial_summary: Option<LocalizedText>

Output only. Contains a summary of the place. A summary is comprised of a textual overview, and also includes the language code for these if applicable. Summary text must be presented as-is and can not be modified or altered.

outdoor_seating: Option<bool>

Output only. Place provides outdoor seating.

live_music: Option<bool>

Output only. Place provides live music.

menu_for_children: Option<bool>

Output only. Place has a children’s menu.

serves_cocktails: Option<bool>

Output only. Place serves cocktails.

serves_dessert: Option<bool>

Output only. Place serves dessert.

serves_coffee: Option<bool>

Output only. Place serves coffee.

good_for_children: Option<bool>

Output only. Place is good for children.

allows_dogs: Option<bool>

Output only. Place allows dogs.

restroom: Option<bool>

Output only. Place has restroom.

good_for_groups: Option<bool>

Output only. Place accommodates groups.

good_for_watching_sports: Option<bool>

Output only. Place is suitable for watching sports.

payment_options: Option<PaymentOptions>parking_options: Option<ParkingOptions>

Output only. Options of parking provided by the place.

sub_destinations: Vec<SubDestination>

Output only. A list of sub destinations related to the place.

accessibility_options: Option<AccessibilityOptions>

Output only. Information about the accessibility options a place offers.

fuel_options: Option<FuelOptions>

Output only. The most recent information about fuel options in a gas station. This information is updated regularly.

ev_charge_options: Option<EvChargeOptions>

Output only. Information of ev charging options.

Implementations

impl Place

pub fn utc_offset_minutes(&self) -> i32

Returns the value of utc_offset_minutes, or the default value if utc_offset_minutes is unset.

pub fn business_status(&self) -> BusinessStatus

Returns the enum value of business_status, or the default if the field is set to an invalid enum value.

pub fn set_business_status(&mut self, value: BusinessStatus)

Sets business_status to the provided enum value.

pub fn price_level(&self) -> PriceLevel

Returns the enum value of price_level, or the default if the field is set to an invalid enum value.

pub fn set_price_level(&mut self, value: PriceLevel)

Sets price_level to the provided enum value.

pub fn user_rating_count(&self) -> i32

Returns the value of user_rating_count, or the default value if user_rating_count is unset.

pub fn takeout(&self) -> bool

Returns the value of takeout, or the default value if takeout is unset.

pub fn delivery(&self) -> bool

Returns the value of delivery, or the default value if delivery is unset.

pub fn dine_in(&self) -> bool

Returns the value of dine_in, or the default value if dine_in is unset.

pub fn curbside_pickup(&self) -> bool

Returns the value of curbside_pickup, or the default value if curbside_pickup is unset.

pub fn reservable(&self) -> bool

Returns the value of reservable, or the default value if reservable is unset.

pub fn serves_breakfast(&self) -> bool

Returns the value of serves_breakfast, or the default value if serves_breakfast is unset.

pub fn serves_lunch(&self) -> bool

Returns the value of serves_lunch, or the default value if serves_lunch is unset.

pub fn serves_dinner(&self) -> bool

Returns the value of serves_dinner, or the default value if serves_dinner is unset.

pub fn serves_beer(&self) -> bool

Returns the value of serves_beer, or the default value if serves_beer is unset.

pub fn serves_wine(&self) -> bool

Returns the value of serves_wine, or the default value if serves_wine is unset.

pub fn serves_brunch(&self) -> bool

Returns the value of serves_brunch, or the default value if serves_brunch is unset.

pub fn serves_vegetarian_food(&self) -> bool

Returns the value of serves_vegetarian_food, or the default value if serves_vegetarian_food is unset.

pub fn outdoor_seating(&self) -> bool

Returns the value of outdoor_seating, or the default value if outdoor_seating is unset.

pub fn live_music(&self) -> bool

Returns the value of live_music, or the default value if live_music is unset.

pub fn menu_for_children(&self) -> bool

Returns the value of menu_for_children, or the default value if menu_for_children is unset.

pub fn serves_cocktails(&self) -> bool

Returns the value of serves_cocktails, or the default value if serves_cocktails is unset.

pub fn serves_dessert(&self) -> bool

Returns the value of serves_dessert, or the default value if serves_dessert is unset.

pub fn serves_coffee(&self) -> bool

Returns the value of serves_coffee, or the default value if serves_coffee is unset.

pub fn good_for_children(&self) -> bool

Returns the value of good_for_children, or the default value if good_for_children is unset.

pub fn allows_dogs(&self) -> bool

Returns the value of allows_dogs, or the default value if allows_dogs is unset.

pub fn restroom(&self) -> bool

Returns the value of restroom, or the default value if restroom is unset.

pub fn good_for_groups(&self) -> bool

Returns the value of good_for_groups, or the default value if good_for_groups is unset.

pub fn good_for_watching_sports(&self) -> bool

Returns the value of good_for_watching_sports, or the default value if good_for_watching_sports is unset.