Module google_api_proto::google::maps::routing::v2

Modules

compute_route_matrix_request

Nested message and enum types in ComputeRouteMatrixRequest.

compute_routes_request

Nested message and enum types in ComputeRoutesRequest.

polyline

Nested message and enum types in Polyline.

route

Nested message and enum types in Route.

route_leg

Nested message and enum types in RouteLeg.

route_leg_step

Nested message and enum types in RouteLegStep.

route_leg_step_transit_details

Nested message and enum types in RouteLegStepTransitDetails.

route_matrix_element

Nested message and enum types in RouteMatrixElement.

routes_client

Generated client implementations.

speed_reading_interval

Nested message and enum types in SpeedReadingInterval.

transit_preferences

Nested message and enum types in TransitPreferences.

transit_vehicle

Nested message and enum types in TransitVehicle.

waypoint

Nested message and enum types in Waypoint.

Structs

ComputeRouteMatrixRequest

ComputeRouteMatrix request message

ComputeRoutesRequest

ComputeRoutes request message.

ComputeRoutesResponse

ComputeRoutes the response message.

FallbackInfo

Information related to how and why a fallback result was used. If this field is set, then it means the server used a different routing mode from your preferred mode as fallback.

GeocodedWaypoint

Details about the locations used as waypoints. Only populated for address waypoints. Includes details about the geocoding results for the purposes of determining what the address was geocoded to.

GeocodingResults

Contains [GeocodedWaypoints][google.maps.routing.v2.GeocodedWaypoint] for origin, destination and intermediate waypoints. Only populated for address waypoints.

LocalizedTime

Localized description of time.

Location

Encapsulates a location (a geographic point, and an optional heading).

NavigationInstruction

Encapsulates navigation instructions for a [RouteLegStep][google.maps.routing.v2.RouteLegStep]

Polyline

Encapsulates an encoded polyline.

Route

Contains a route, which consists of a series of connected road segments that join beginning, ending, and intermediate waypoints.

RouteLeg

Contains a segment between non-via waypoints.

RouteLegStep

Contains a segment of a [RouteLeg][google.maps.routing.v2.RouteLeg]. A step corresponds to a single navigation instruction. Route legs are made up of steps.

RouteLegStepTransitDetails

Additional information for the RouteLegStep related to TRANSIT routes.

RouteLegStepTravelAdvisory

Contains the additional information that the user should be informed about, such as possible traffic zone restrictions on a leg step.

RouteLegTravelAdvisory

Contains the additional information that the user should be informed about on a leg step, such as possible traffic zone restrictions.

RouteMatrixDestination

A single destination for ComputeRouteMatrixRequest

RouteMatrixElement

Contains route information computed for an origin/destination pair in the ComputeRouteMatrix API. This proto can be streamed to the client.

RouteMatrixOrigin

A single origin for ComputeRouteMatrixRequest

RouteModifiers

Encapsulates a set of optional conditions to satisfy when calculating the routes.

RouteTravelAdvisory

Contains the additional information that the user should be informed about, such as possible traffic zone restrictions.

SpeedReadingInterval

Traffic density indicator on a contiguous segment of a polyline or path. Given a path with points P_0, P_1, … , P_N (zero-based index), the SpeedReadingInterval defines an interval and describes its traffic using the following categories.

TollInfo

Encapsulates toll information on a [Route][google.maps.routing.v2.Route] or on a [RouteLeg][google.maps.routing.v2.RouteLeg].

TransitAgency

A transit agency that operates a transit line.

TransitLine

Contains information about the transit line used in this step.

TransitPreferences

Preferences for TRANSIT based routes that influence the route that is returned.

TransitStop

Information about a transit stop.

TransitVehicle

Information about a vehicle used in transit routes.

VehicleInfo

Contains the vehicle information, such as the vehicle emission type.

Waypoint

Encapsulates a waypoint. Waypoints mark both the beginning and end of a route, and include intermediate stops along the route.

Enums

FallbackReason

Reasons for using fallback response.

FallbackRoutingMode

Actual routing mode used for returned fallback response.

Maneuver

A set of values that specify the navigation action to take for the current step (e.g., turn left, merge, straight, etc.).

PolylineEncoding

Specifies the preferred type of polyline to be returned.

PolylineQuality

A set of values that specify the quality of the polyline.

RouteLabel

Labels for the [Route][google.maps.routing.v2.Route] that are useful to identify specific properties of the route to compare against others.

RouteMatrixElementCondition

The condition of the route being returned.

RouteTravelMode

A set of values used to specify the mode of travel. NOTE: WALK, BICYCLE, and TWO_WHEELER routes are in beta and might sometimes be missing clear sidewalks, pedestrian paths, or bicycling paths. You must display this warning to the user for all walking, bicycling, and two-wheel routes that you display in your app.

RoutingPreference

A set of values that specify factors to take into consideration when calculating the route.

TollPass

List of toll passes around the world that we support.

TrafficModel

Specifies the assumptions to use when calculating time in traffic. This setting affects the value returned in the duration field in the response, which contains the predicted time in traffic based on historical averages.

Units

A set of values that specify the unit of measure used in the display.

VehicleEmissionType

A set of values describing the vehicle’s emission type. Applies only to the DRIVE [RouteTravelMode][google.maps.routing.v2.RouteTravelMode].