Enum google_api_proto::google::maps::routing::v2::TollPass

#[repr(i32)]
pub enum TollPass {

    Unspecified = 0,
    AuEtollTag = 82,
    AuEwayTag = 83,
    AuLinkt = 2,
    ArTelepase = 3,
    BrAutoExpreso = 81,
    BrConectcar = 7,
    BrMoveMais = 8,
    BrPassaRapido = 88,
    BrSemParar = 9,
    BrTaggy = 10,
    BrVeloe = 11,
    CaUsAkwasasneSeawayCorporateCard = 84,
    CaUsAkwasasneSeawayTransitCard = 85,
    CaUsBlueWaterEdgePass = 18,
    CaUsConnexion = 19,
    CaUsNexusCard = 20,
    IdEToll = 16,
    InFastag = 78,
    InLocalHpPlateExempt = 79,
    MxIave = 90,
    MxPase = 91,
    MxQuickpass = 93,
    MxSistemaTelepeajeChihuahua = 89,
    MxTagIave = 12,
    MxTagTelevia = 13,
    MxTelevia = 92,
    MxViapass = 14,
    UsAlFreedomPass = 21,
    UsAkAntonAndersonTunnelBookOf10Tickets = 22,
    UsCaFastrak = 4,
    UsCaFastrakCavSticker = 86,
    UsCoExpresstoll = 23,
    UsCoGoPass = 24,
    UsDeEzpassde = 25,
    UsFlBobSikesTollBridgePass = 65,
    UsFlDunesCommunityDevelopmentDistrictExpresscard = 66,
    UsFlEpass = 67,
    UsFlGibaTollPass = 68,
    UsFlLeeway = 69,
    UsFlSunpass = 70,
    UsFlSunpassPro = 71,
    UsIlEzpassil = 73,
    UsIlIpass = 72,
    UsInEzpassin = 26,
    UsKsBestpassHorizon = 27,
    UsKsKtag = 28,
    UsKsNationalpass = 29,
    UsKsPrepassElitepass = 30,
    UsKyRiverlink = 31,
    UsLaGeauxpass = 32,
    UsLaTollTag = 33,
    UsMaEzpassma = 6,
    UsMdEzpassmd = 34,
    UsMeEzpassme = 35,
    UsMiAmbassadorBridgePremierCommuterCard = 36,
    UsMiBcpass = 94,
    UsMiGrosseIleTollBridgePassTag = 37,
    UsMiIqProxCard = 38,
    UsMiIqTag = 95,
    UsMiMackinacBridgeMacPass = 39,
    UsMiNexpressToll = 40,
    UsMnEzpassmn = 41,
    UsNcEzpassnc = 42,
    UsNcPeachPass = 87,
    UsNcQuickPass = 43,
    UsNhEzpassnh = 80,
    UsNjDownbeachExpressPass = 75,
    UsNjEzpassnj = 74,
    UsNyExpresspass = 76,
    UsNyEzpassny = 77,
    UsOhEzpassoh = 44,
    UsPaEzpasspa = 45,
    UsRiEzpassri = 46,
    UsScPalpass = 47,
    UsTxAviTag = 97,
    UsTxBancpass = 48,
    UsTxDelRioPass = 49,
    UsTxEfastPass = 50,
    UsTxEaglePassExpressCard = 51,
    UsTxEptoll = 52,
    UsTxEzCross = 53,
    UsTxEztag = 54,
    UsTxFuegoTag = 96,
    UsTxLaredoTradeTag = 55,
    UsTxPluspass = 56,
    UsTxTolltag = 57,
    UsTxTxtag = 58,
    UsTxXpressCard = 59,
    UsUtAdamsAveParkwayExpresscard = 60,
    UsVaEzpassva = 61,
    UsWaBreezeby = 17,
    UsWaGoodToGo = 1,
    UsWvEzpasswv = 62,
    UsWvMemorialBridgeTickets = 63,
    UsWvNewellTollBridgeTicket = 64,

}

List of toll passes around the world that we support.

Variants

Unspecified = 0

Not used. If this value is used, then the request fails.

AuEtollTag = 82

Sydney toll pass. See additional details at https://www.myetoll.com.au.

AuEwayTag = 83

Sydney toll pass. See additional details at https://www.tollpay.com.au.

AuLinkt = 2

Australia-wide toll pass. See additional details at https://www.linkt.com.au/.

ArTelepase = 3

Argentina toll pass. See additional details at https://telepase.com.ar

BrAutoExpreso = 81

Brazil toll pass. See additional details at https://www.autoexpreso.com

BrConectcar = 7

Brazil toll pass. See additional details at https://conectcar.com.

BrMoveMais = 8

Brazil toll pass. See additional details at https://movemais.com.

BrPassaRapido = 88

Brazil toll pass. See additional details at https://pasorapido.gob.do/

BrSemParar = 9

Brazil toll pass. See additional details at https://www.semparar.com.br.

BrTaggy = 10

Brazil toll pass. See additional details at https://taggy.com.br.

BrVeloe = 11

Brazil toll pass. See additional details at https://veloe.com.br/site/onde-usar.

CaUsAkwasasneSeawayCorporateCard = 84

Canada to United States border crossing.

CaUsAkwasasneSeawayTransitCard = 85

Canada to United States border crossing.

CaUsBlueWaterEdgePass = 18

Ontario, Canada to Michigan, United States border crossing.

CaUsConnexion = 19

Ontario, Canada to Michigan, United States border crossing.

CaUsNexusCard = 20

Canada to United States border crossing.

IdEToll = 16

Indonesia. E-card provided by multiple banks used to pay for tolls. All e-cards via banks are charged the same so only one enum value is needed. E.g.

Bank Mandiri https://www.bankmandiri.co.id/e-money

BCA https://www.bca.co.id/flazz

BNI https://www.bni.co.id/id-id/ebanking/tapcash

InFastag = 78

India.

InLocalHpPlateExempt = 79

India, HP state plate exemption.

MxIave = 90

Mexico toll pass. https://iave.capufe.gob.mx/#/

MxPase = 91

Mexico https://www.pase.com.mx

MxQuickpass = 93

Mexico https://operadoravial.com/quick-pass/

MxSistemaTelepeajeChihuahua = 89

http://appsh.chihuahua.gob.mx/transparencia/?doc=/ingresos/TelepeajeFormato4.pdf

MxTagIave = 12

Mexico

MxTagTelevia = 13

Mexico toll pass company. One of many operating in Mexico City. See additional details at https://www.televia.com.mx.

MxTelevia = 92

Mexico toll pass company. One of many operating in Mexico City. https://www.televia.com.mx

MxViapass = 14

Mexico toll pass. See additional details at https://www.viapass.com.mx/viapass/web_home.aspx.

UsAlFreedomPass = 21

AL, USA.

UsAkAntonAndersonTunnelBookOf10Tickets = 22

AK, USA.

UsCaFastrak = 4

CA, USA.

UsCaFastrakCavSticker = 86

Indicates driver has any FasTrak pass in addition to the DMV issued Clean Air Vehicle (CAV) sticker. https://www.bayareafastrak.org/en/guide/doINeedFlex.shtml

UsCoExpresstoll = 23

CO, USA.

UsCoGoPass = 24

CO, USA.

UsDeEzpassde = 25

DE, USA.

UsFlBobSikesTollBridgePass = 65

FL, USA.

UsFlDunesCommunityDevelopmentDistrictExpresscard = 66

FL, USA.

UsFlEpass = 67

FL, USA.

UsFlGibaTollPass = 68

FL, USA.

UsFlLeeway = 69

FL, USA.

UsFlSunpass = 70

FL, USA.

UsFlSunpassPro = 71

FL, USA.

UsIlEzpassil = 73

IL, USA.

UsIlIpass = 72

IL, USA.

UsInEzpassin = 26

IN, USA.

UsKsBestpassHorizon = 27

KS, USA.

UsKsKtag = 28

KS, USA.

UsKsNationalpass = 29

KS, USA.

UsKsPrepassElitepass = 30

KS, USA.

UsKyRiverlink = 31

KY, USA.

UsLaGeauxpass = 32

LA, USA.

UsLaTollTag = 33

LA, USA.

UsMaEzpassma = 6

MA, USA.

UsMdEzpassmd = 34

MD, USA.

UsMeEzpassme = 35

ME, USA.

UsMiAmbassadorBridgePremierCommuterCard = 36

MI, USA.

UsMiBcpass = 94

MI, USA.

UsMiGrosseIleTollBridgePassTag = 37

MI, USA.

UsMiIqProxCard = 38

MI, USA. Deprecated as this pass type no longer exists.

UsMiIqTag = 95

MI, USA.

UsMiMackinacBridgeMacPass = 39

MI, USA.

UsMiNexpressToll = 40

MI, USA.

UsMnEzpassmn = 41

MN, USA.

UsNcEzpassnc = 42

NC, USA.

UsNcPeachPass = 87

NC, USA.

UsNcQuickPass = 43

NC, USA.

UsNhEzpassnh = 80

NH, USA.

UsNjDownbeachExpressPass = 75

NJ, USA.

UsNjEzpassnj = 74

NJ, USA.

UsNyExpresspass = 76

NY, USA.

UsNyEzpassny = 77

NY, USA.

UsOhEzpassoh = 44

OH, USA.

UsPaEzpasspa = 45

PA, USA.

UsRiEzpassri = 46

RI, USA.

UsScPalpass = 47

SC, USA.

UsTxAviTag = 97

TX, USA.

UsTxBancpass = 48

TX, USA.

UsTxDelRioPass = 49

TX, USA.

UsTxEfastPass = 50

TX, USA.

UsTxEaglePassExpressCard = 51

TX, USA.

UsTxEptoll = 52

TX, USA.

UsTxEzCross = 53

TX, USA.

UsTxEztag = 54

TX, USA.

UsTxFuegoTag = 96

TX, USA.

UsTxLaredoTradeTag = 55

TX, USA.

UsTxPluspass = 56

TX, USA.

UsTxTolltag = 57

TX, USA.

UsTxTxtag = 58

TX, USA.

UsTxXpressCard = 59

TX, USA.

UsUtAdamsAveParkwayExpresscard = 60

UT, USA.

UsVaEzpassva = 61

VA, USA.

UsWaBreezeby = 17

WA, USA.

UsWaGoodToGo = 1

WA, USA.

UsWvEzpasswv = 62

WV, USA.

UsWvMemorialBridgeTickets = 63

WV, USA.

UsWvNewellTollBridgeTicket = 64

WV, USA.

Implementations

impl TollPass

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of TollPass.

pub fn from_i32(value: i32) -> Option<TollPass>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a TollPass, or None if value is not a valid variant.

impl TollPass

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.