Module google_api_proto::google::monitoring::v3::alert_policy::condition

Nested message and enum types in Condition.

Modules

metric_threshold

Nested message and enum types in MetricThreshold.

trigger

Nested message and enum types in Trigger.

Structs

LogMatch

A condition type that checks whether a log message in the scoping project satisfies the given filter. Logs from other projects in the metrics scope are not evaluated.

MetricAbsence

A condition type that checks that monitored resources are reporting data. The configuration defines a metric and a set of monitored resources. The predicate is considered in violation when a time series for the specified metric of a monitored resource does not include any data in the specified duration.

MetricThreshold

A condition type that compares a collection of time series against a threshold.

MonitoringQueryLanguageCondition

A condition type that allows alert policies to be defined using Monitoring Query Language.

PrometheusQueryLanguageCondition

A condition type that allows alert policies to be defined using Prometheus Query Language (PromQL).

Trigger

Specifies how many time series must fail a predicate to trigger a condition. If not specified, then a {count: 1} trigger is used.

Enums

Condition

Only one of the following condition types will be specified.

EvaluationMissingData

A condition control that determines how metric-threshold conditions are evaluated when data stops arriving. This control doesn’t affect metric-absence policies.