Struct google_api_proto::google::monitoring::v3::AlertPolicy

pub struct AlertPolicy {

    pub name: String,
    pub display_name: String,
    pub documentation: Option<Documentation>,
    pub user_labels: BTreeMap<String, String>,
    pub conditions: Vec<Condition>,
    pub combiner: i32,
    pub enabled: Option<bool>,
    pub validity: Option<Status>,
    pub notification_channels: Vec<String>,
    pub creation_record: Option<MutationRecord>,
    pub mutation_record: Option<MutationRecord>,
    pub alert_strategy: Option<AlertStrategy>,
    pub severity: i32,

}

A description of the conditions under which some aspect of your system is considered to be “unhealthy” and the ways to notify people or services about this state. For an overview of alert policies, see Introduction to Alerting.

Fields
name: String

Required if the policy exists. The resource name for this policy. The format is:

 projects/\[PROJECT_ID_OR_NUMBER\]/alertPolicies/\[ALERT_POLICY_ID\]

\[ALERT_POLICY_ID\] is assigned by Cloud Monitoring when the policy is created. When calling the [alertPolicies.create][google.monitoring.v3.AlertPolicyService.CreateAlertPolicy] method, do not include the name field in the alerting policy passed as part of the request.

display_name: String

A short name or phrase used to identify the policy in dashboards, notifications, and incidents. To avoid confusion, don’t use the same display name for multiple policies in the same project. The name is limited to 512 Unicode characters.

The convention for the display_name of a PrometheusQueryLanguageCondition is “{rule group name}/{alert name}”, where the {rule group name} and {alert name} should be taken from the corresponding Prometheus configuration file. This convention is not enforced. In any case the display_name is not a unique key of the AlertPolicy.

documentation: Option<Documentation>

Documentation that is included with notifications and incidents related to this policy. Best practice is for the documentation to include information to help responders understand, mitigate, escalate, and correct the underlying problems detected by the alerting policy. Notification channels that have limited capacity might not show this documentation.

user_labels: BTreeMap<String, String>

User-supplied key/value data to be used for organizing and identifying the AlertPolicy objects.

The field can contain up to 64 entries. Each key and value is limited to 63 Unicode characters or 128 bytes, whichever is smaller. Labels and values can contain only lowercase letters, numerals, underscores, and dashes. Keys must begin with a letter.

Note that Prometheus {alert name} is a valid Prometheus label names, whereas Prometheus {rule group} is an unrestricted UTF-8 string. This means that they cannot be stored as-is in user labels, because they may contain characters that are not allowed in user-label values.

conditions: Vec<Condition>

A list of conditions for the policy. The conditions are combined by AND or OR according to the combiner field. If the combined conditions evaluate to true, then an incident is created. A policy can have from one to six conditions. If condition_time_series_query_language is present, it must be the only condition. If condition_monitoring_query_language is present, it must be the only condition.

combiner: i32

How to combine the results of multiple conditions to determine if an incident should be opened. If condition_time_series_query_language is present, this must be COMBINE_UNSPECIFIED.

enabled: Option<bool>

Whether or not the policy is enabled. On write, the default interpretation if unset is that the policy is enabled. On read, clients should not make any assumption about the state if it has not been populated. The field should always be populated on List and Get operations, unless a field projection has been specified that strips it out.

validity: Option<Status>

Read-only description of how the alert policy is invalid. This field is only set when the alert policy is invalid. An invalid alert policy will not generate incidents.

notification_channels: Vec<String>

Identifies the notification channels to which notifications should be sent when incidents are opened or closed or when new violations occur on an already opened incident. Each element of this array corresponds to the name field in each of the [NotificationChannel][google.monitoring.v3.NotificationChannel] objects that are returned from the [ListNotificationChannels] [google.monitoring.v3.NotificationChannelService.ListNotificationChannels] method. The format of the entries in this field is:

 projects/\[PROJECT_ID_OR_NUMBER\]/notificationChannels/\[CHANNEL_ID\]

creation_record: Option<MutationRecord>

A read-only record of the creation of the alerting policy. If provided in a call to create or update, this field will be ignored.

mutation_record: Option<MutationRecord>

A read-only record of the most recent change to the alerting policy. If provided in a call to create or update, this field will be ignored.

alert_strategy: Option<AlertStrategy>

Control over how this alert policy’s notification channels are notified.

severity: i32

Optional. The severity of an alert policy indicates how important incidents generated by that policy are. The severity level will be displayed on the Incident detail page and in notifications.

Implementations

impl AlertPolicy

pub fn combiner(&self) -> ConditionCombinerType

Returns the enum value of combiner, or the default if the field is set to an invalid enum value.

pub fn set_combiner(&mut self, value: ConditionCombinerType)

Sets combiner to the provided enum value.

pub fn severity(&self) -> Severity

Returns the enum value of severity, or the default if the field is set to an invalid enum value.

pub fn set_severity(&mut self, value: Severity)

Sets severity to the provided enum value.