Module google_api_proto::google::privacy::dlp::v2

Modules

action

Nested message and enum types in Action.

action_details

Nested message and enum types in ActionDetails.

analyze_data_source_risk_details

Nested message and enum types in AnalyzeDataSourceRiskDetails.

big_query_discovery_target

Nested message and enum types in BigQueryDiscoveryTarget.

big_query_options

Nested message and enum types in BigQueryOptions.

big_query_table_collection

Nested message and enum types in BigQueryTableCollection.

bucketing_config

Nested message and enum types in BucketingConfig.

byte_content_item

Nested message and enum types in ByteContentItem.

chars_to_ignore

Nested message and enum types in CharsToIgnore.

cloud_storage_options

Nested message and enum types in CloudStorageOptions.

column_data_profile

Nested message and enum types in ColumnDataProfile.

content_item

Nested message and enum types in ContentItem.

content_location

Nested message and enum types in ContentLocation.

create_dlp_job_request

Nested message and enum types in CreateDlpJobRequest.

crypto_key

Nested message and enum types in CryptoKey.

crypto_replace_ffx_fpe_config

Nested message and enum types in CryptoReplaceFfxFpeConfig.

custom_info_type

Nested message and enum types in CustomInfoType.

data_profile_action

Nested message and enum types in DataProfileAction.

data_profile_big_query_row_schema

Nested message and enum types in DataProfileBigQueryRowSchema.

data_profile_location

Nested message and enum types in DataProfileLocation.

data_profile_pub_sub_condition

Nested message and enum types in DataProfilePubSubCondition.

data_risk_level

Nested message and enum types in DataRiskLevel.

date_shift_config

Nested message and enum types in DateShiftConfig.

date_time

Nested message and enum types in DateTime.

deidentify_config

Nested message and enum types in DeidentifyConfig.

deidentify_data_source_details

Nested message and enum types in DeidentifyDataSourceDetails.

discovery_big_query_conditions

Nested message and enum types in DiscoveryBigQueryConditions.

discovery_big_query_filter

Nested message and enum types in DiscoveryBigQueryFilter.

discovery_config

Nested message and enum types in DiscoveryConfig.

discovery_starting_location

Nested message and enum types in DiscoveryStartingLocation.

discovery_target

Nested message and enum types in DiscoveryTarget.

dlp_job

Nested message and enum types in DlpJob.

dlp_service_client

Generated client implementations.

exclusion_rule

Nested message and enum types in ExclusionRule.

field_transformation

Nested message and enum types in FieldTransformation.

image_transformations

Nested message and enum types in ImageTransformations.

info_type_category

Nested message and enum types in InfoTypeCategory.

info_type_transformations

Nested message and enum types in InfoTypeTransformations.

inspect_config

Nested message and enum types in InspectConfig.

inspect_data_source_details

Nested message and enum types in InspectDataSourceDetails.

inspection_rule

Nested message and enum types in InspectionRule.

job_trigger

Nested message and enum types in JobTrigger.

key

Nested message and enum types in Key.

large_custom_dictionary_config

Nested message and enum types in LargeCustomDictionaryConfig.

metadata_location

Nested message and enum types in MetadataLocation.

output_storage_config

Nested message and enum types in OutputStorageConfig.

primitive_transformation

Nested message and enum types in PrimitiveTransformation.

privacy_metric

Nested message and enum types in PrivacyMetric.

quasi_id

Nested message and enum types in QuasiId.

quote_info

Nested message and enum types in QuoteInfo.

record_condition

Nested message and enum types in RecordCondition.

record_key

Nested message and enum types in RecordKey.

redact_image_request

Nested message and enum types in RedactImageRequest.

replace_dictionary_config

Nested message and enum types in ReplaceDictionaryConfig.

schedule

Nested message and enum types in Schedule.

sensitivity_score

Nested message and enum types in SensitivityScore.

statistical_table

Nested message and enum types in StatisticalTable.

storage_config

Nested message and enum types in StorageConfig.

stored_info_type_config

Nested message and enum types in StoredInfoTypeConfig.

stored_info_type_stats

Nested message and enum types in StoredInfoTypeStats.

table

Nested message and enum types in Table.

table_data_profile

Nested message and enum types in TableDataProfile.

time_part_config

Nested message and enum types in TimePartConfig.

transformation_details_storage_config

Nested message and enum types in TransformationDetailsStorageConfig.

transformation_error_handling

Nested message and enum types in TransformationErrorHandling.

transformation_location

Nested message and enum types in TransformationLocation.

transformation_summary

Nested message and enum types in TransformationSummary.

value

Nested message and enum types in Value.

Structs

Action

A task to execute on the completion of a job. See https://cloud.google.com/dlp/docs/concepts-actions to learn more.

ActionDetails

The results of an [Action][google.privacy.dlp.v2.Action].

ActivateJobTriggerRequest

Request message for ActivateJobTrigger.

AnalyzeDataSourceRiskDetails

Result of a risk analysis operation request.

BigQueryDiscoveryTarget

Target used to match against for discovery with BigQuery tables

BigQueryField

Message defining a field of a BigQuery table.

BigQueryKey

Row key for identifying a record in BigQuery table.

BigQueryOptions

Options defining BigQuery table and row identifiers.

BigQueryRegex

A pattern to match against one or more tables, datasets, or projects that contain BigQuery tables. At least one pattern must be specified. Regular expressions use RE2 syntax; a guide can be found under the google/re2 repository on GitHub.

BigQueryRegexes

A collection of regular expressions to determine what tables to match against.

BigQueryTable

Message defining the location of a BigQuery table. A table is uniquely identified by its project_id, dataset_id, and table_name. Within a query a table is often referenced with a string in the format of: <project_id>:<dataset_id>.<table_id> or <project_id>.<dataset_id>.<table_id>.

BigQueryTableCollection

Specifies a collection of BigQuery tables. Used for Discovery.

BigQueryTableTypes

The types of BigQuery tables supported by Cloud DLP.

BoundingBox

Bounding box encompassing detected text within an image.

BucketingConfig

Generalization function that buckets values based on ranges. The ranges and replacement values are dynamically provided by the user for custom behavior, such as 1-30 -> LOW 31-65 -> MEDIUM 66-100 -> HIGH This can be used on data of type: number, long, string, timestamp. If the bound Value type differs from the type of data being transformed, we will first attempt converting the type of the data to be transformed to match the type of the bound before comparing. See https://cloud.google.com/dlp/docs/concepts-bucketing to learn more.

ByteContentItem

Container for bytes to inspect or redact.

CancelDlpJobRequest

The request message for canceling a DLP job.

CharacterMaskConfig

Partially mask a string by replacing a given number of characters with a fixed character. Masking can start from the beginning or end of the string. This can be used on data of any type (numbers, longs, and so on) and when de-identifying structured data we’ll attempt to preserve the original data’s type. (This allows you to take a long like 123 and modify it to a string like **3.

CharsToIgnore

Characters to skip when doing deidentification of a value. These will be left alone and skipped.

CloudStorageFileSet

Message representing a set of files in Cloud Storage.

CloudStorageOptions

Options defining a file or a set of files within a Cloud Storage bucket.

CloudStoragePath

Message representing a single file or path in Cloud Storage.

CloudStorageRegexFileSet

Message representing a set of files in a Cloud Storage bucket. Regular expressions are used to allow fine-grained control over which files in the bucket to include.

Color

Represents a color in the RGB color space.

ColumnDataProfile

The profile for a scanned column within a table.

Container

Represents a container that may contain DLP findings. Examples of a container include a file, table, or database record.

ContentItem

ContentLocation

Precise location of the finding within a document, record, image, or metadata container.

CreateDeidentifyTemplateRequest

Request message for CreateDeidentifyTemplate.

CreateDiscoveryConfigRequest

Request message for CreateDiscoveryConfig.

CreateDlpJobRequest

Request message for CreateDlpJobRequest. Used to initiate long running jobs such as calculating risk metrics or inspecting Google Cloud Storage.

CreateInspectTemplateRequest

Request message for CreateInspectTemplate.

CreateJobTriggerRequest

Request message for CreateJobTrigger.

CreateStoredInfoTypeRequest

Request message for CreateStoredInfoType.

CryptoDeterministicConfig

Pseudonymization method that generates deterministic encryption for the given input. Outputs a base64 encoded representation of the encrypted output. Uses AES-SIV based on the RFC https://tools.ietf.org/html/rfc5297.

CryptoHashConfig

Pseudonymization method that generates surrogates via cryptographic hashing. Uses SHA-256. The key size must be either 32 or 64 bytes. Outputs a base64 encoded representation of the hashed output (for example, L7k0BHmF1ha5U3NfGykjro4xWi1MPVQPjhMAZbSV9mM=). Currently, only string and integer values can be hashed. See https://cloud.google.com/dlp/docs/pseudonymization to learn more.

CryptoKey

This is a data encryption key (DEK) (as opposed to a key encryption key (KEK) stored by Cloud Key Management Service (Cloud KMS). When using Cloud KMS to wrap or unwrap a DEK, be sure to set an appropriate IAM policy on the KEK to ensure an attacker cannot unwrap the DEK.

CryptoReplaceFfxFpeConfig

Replaces an identifier with a surrogate using Format Preserving Encryption (FPE) with the FFX mode of operation; however when used in the ReidentifyContent API method, it serves the opposite function by reversing the surrogate back into the original identifier. The identifier must be encoded as ASCII. For a given crypto key and context, the same identifier will be replaced with the same surrogate. Identifiers must be at least two characters long. In the case that the identifier is the empty string, it will be skipped. See https://cloud.google.com/dlp/docs/pseudonymization to learn more.

CustomInfoType

Custom information type provided by the user. Used to find domain-specific sensitive information configurable to the data in question.

DataProfileAction

A task to execute when a data profile has been generated.

DataProfileBigQueryRowSchema

The schema of data to be saved to the BigQuery table when the DataProfileAction is enabled.

DataProfileConfigSnapshot

Snapshot of the configurations used to generate the profile.

DataProfileJobConfig

Configuration for setting up a job to scan resources for profile generation. Only one data profile configuration may exist per organization, folder, or project.

DataProfileLocation

The data that will be profiled.

DataProfilePubSubCondition

A condition for determining whether a Pub/Sub should be triggered.

DataProfilePubSubMessage

Pub/Sub topic message for a DataProfileAction.PubSubNotification event. To receive a message of protocol buffer schema type, convert the message data to an object of this proto class.

DataRiskLevel

Score is a summary of all elements in the data profile. A higher number means more risk.

DatastoreKey

Record key for a finding in Cloud Datastore.

DatastoreOptions

Options defining a data set within Google Cloud Datastore.

DateShiftConfig

Shifts dates by random number of days, with option to be consistent for the same context. See https://cloud.google.com/dlp/docs/concepts-date-shifting to learn more.

DateTime

Message for a date time object. e.g. 2018-01-01, 5th August.

DeidentifyConfig

The configuration that controls how the data will change.

DeidentifyContentRequest

Request to de-identify a ContentItem.

DeidentifyContentResponse

Results of de-identifying a ContentItem.

DeidentifyDataSourceDetails

The results of a [Deidentify][google.privacy.dlp.v2.Action.Deidentify] action from an inspect job.

DeidentifyDataSourceStats

Summary of what was modified during a transformation.

DeidentifyTemplate

DeidentifyTemplates contains instructions on how to de-identify content. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

DeleteDeidentifyTemplateRequest

Request message for DeleteDeidentifyTemplate.

DeleteDiscoveryConfigRequest

Request message for DeleteDiscoveryConfig.

DeleteDlpJobRequest

The request message for deleting a DLP job.

DeleteInspectTemplateRequest

Request message for DeleteInspectTemplate.

DeleteJobTriggerRequest

Request message for DeleteJobTrigger.

DeleteStoredInfoTypeRequest

Request message for DeleteStoredInfoType.

Disabled

Do not profile the tables.

DiscoveryBigQueryConditions

Requirements that must be true before a table is scanned in discovery for the first time. There is an AND relationship between the top-level attributes. Additionally, minimum conditions with an OR relationship that must be met before Cloud DLP scans a table can be set (like a minimum row count or a minimum table age).

DiscoveryBigQueryFilter

Determines what tables will have profiles generated within an organization or project. Includes the ability to filter by regular expression patterns on project ID, dataset ID, and table ID.

DiscoveryConfig

Configuration for discovery to scan resources for profile generation. Only one discovery configuration may exist per organization, folder, or project.

DiscoveryGenerationCadence

What must take place for a profile to be updated and how frequently it should occur. New tables are scanned as quickly as possible depending on system capacity.

DiscoverySchemaModifiedCadence

The cadence at which to update data profiles when a schema is modified.

DiscoveryStartingLocation

The location to begin a discovery scan. Denotes an organization ID or folder ID within an organization.

DiscoveryTableModifiedCadence

The cadence at which to update data profiles when a table is modified.

DiscoveryTarget

Target used to match against for Discovery.

DlpJob

Combines all of the information about a DLP job.

DocumentLocation

Location of a finding within a document.

EntityId

An entity in a dataset is a field or set of fields that correspond to a single person. For example, in medical records the EntityId might be a patient identifier, or for financial records it might be an account identifier. This message is used when generalizations or analysis must take into account that multiple rows correspond to the same entity.

Error

Details information about an error encountered during job execution or the results of an unsuccessful activation of the JobTrigger.

ExcludeByHotword

The rule to exclude findings based on a hotword. For record inspection of tables, column names are considered hotwords. An example of this is to exclude a finding if it belongs to a BigQuery column that matches a specific pattern.

ExcludeInfoTypes

List of excluded infoTypes.

ExclusionRule

The rule that specifies conditions when findings of infoTypes specified in InspectionRuleSet are removed from results.

FieldId

General identifier of a data field in a storage service.

FieldTransformation

The transformation to apply to the field.

Finding

Represents a piece of potentially sensitive content.

FinishDlpJobRequest

The request message for finishing a DLP hybrid job.

FixedSizeBucketingConfig

Buckets values based on fixed size ranges. The Bucketing transformation can provide all of this functionality, but requires more configuration. This message is provided as a convenience to the user for simple bucketing strategies.

GetDeidentifyTemplateRequest

Request message for GetDeidentifyTemplate.

GetDiscoveryConfigRequest

Request message for GetDiscoveryConfig.

GetDlpJobRequest

The request message for [DlpJobs.GetDlpJob][].

GetInspectTemplateRequest

Request message for GetInspectTemplate.

GetJobTriggerRequest

Request message for GetJobTrigger.

GetStoredInfoTypeRequest

Request message for GetStoredInfoType.

HybridContentItem

An individual hybrid item to inspect. Will be stored temporarily during processing.

HybridFindingDetails

Populate to associate additional data with each finding.

HybridInspectDlpJobRequest

Request to search for potentially sensitive info in a custom location.

HybridInspectJobTriggerRequest

Request to search for potentially sensitive info in a custom location.

HybridInspectResponse

Quota exceeded errors will be thrown once quota has been met.

HybridInspectStatistics

Statistics related to processing hybrid inspect requests.

HybridOptions

Configuration to control jobs where the content being inspected is outside of Google Cloud Platform.

ImageLocation

Location of the finding within an image.

ImageTransformations

A type of transformation that is applied over images.

InfoType

Type of information detected by the API.

InfoTypeCategory

Classification of infoTypes to organize them according to geographic location, industry, and data type.

InfoTypeDescription

InfoType description.

InfoTypeStats

Statistics regarding a specific InfoType.

InfoTypeSummary

The infoType details for this column.

InfoTypeTransformations

A type of transformation that will scan unstructured text and apply various PrimitiveTransformations to each finding, where the transformation is applied to only values that were identified as a specific info_type.

InspectConfig

Configuration description of the scanning process. When used with redactContent only info_types and min_likelihood are currently used.

InspectContentRequest

Request to search for potentially sensitive info in a ContentItem.

InspectContentResponse

Results of inspecting an item.

InspectDataSourceDetails

The results of an inspect DataSource job.

InspectJobConfig

Controls what and how to inspect for findings.

InspectResult

All the findings for a single scanned item.

InspectTemplate

The inspectTemplate contains a configuration (set of types of sensitive data to be detected) to be used anywhere you otherwise would normally specify InspectConfig. See https://cloud.google.com/dlp/docs/concepts-templates to learn more.

InspectionRule

A single inspection rule to be applied to infoTypes, specified in InspectionRuleSet.

InspectionRuleSet

Rule set for modifying a set of infoTypes to alter behavior under certain circumstances, depending on the specific details of the rules within the set.

JobTrigger

Contains a configuration to make dlp api calls on a repeating basis. See https://cloud.google.com/dlp/docs/concepts-job-triggers to learn more.

Key

A unique identifier for a Datastore entity. If a key’s partition ID or any of its path kinds or names are reserved/read-only, the key is reserved/read-only. A reserved/read-only key is forbidden in certain documented contexts.

KindExpression

A representation of a Datastore kind.

KmsWrappedCryptoKey

Include to use an existing data crypto key wrapped by KMS. The wrapped key must be a 128-, 192-, or 256-bit key. Authorization requires the following IAM permissions when sending a request to perform a crypto transformation using a KMS-wrapped crypto key: dlp.kms.encrypt

LargeCustomDictionaryConfig

Configuration for a custom dictionary created from a data source of any size up to the maximum size defined in the limits page. The artifacts of dictionary creation are stored in the specified Cloud Storage location. Consider using CustomInfoType.Dictionary for smaller dictionaries that satisfy the size requirements.

LargeCustomDictionaryStats

Summary statistics of a custom dictionary.

ListDeidentifyTemplatesRequest

Request message for ListDeidentifyTemplates.

ListDeidentifyTemplatesResponse

Response message for ListDeidentifyTemplates.

ListDiscoveryConfigsRequest

Request message for ListDiscoveryConfigs.

ListDiscoveryConfigsResponse

Response message for ListDiscoveryConfigs.

ListDlpJobsRequest

The request message for listing DLP jobs.

ListDlpJobsResponse

The response message for listing DLP jobs.

ListInfoTypesRequest

Request for the list of infoTypes.

ListInfoTypesResponse

Response to the ListInfoTypes request.

ListInspectTemplatesRequest

Request message for ListInspectTemplates.

ListInspectTemplatesResponse

Response message for ListInspectTemplates.

ListJobTriggersRequest

Request message for ListJobTriggers.

ListJobTriggersResponse

Response message for ListJobTriggers.

ListStoredInfoTypesRequest

Request message for ListStoredInfoTypes.

ListStoredInfoTypesResponse

Response message for ListStoredInfoTypes.

Location

Specifies the location of the finding.

Manual

Job trigger option for hybrid jobs. Jobs must be manually created and finished.

MetadataLocation

Metadata Location

OtherInfoTypeSummary

Infotype details for other infoTypes found within a column.

OutputStorageConfig

Cloud repository for storing output.

PartitionId

Datastore partition ID. A partition ID identifies a grouping of entities. The grouping is always by project and namespace, however the namespace ID may be empty.

PrimitiveTransformation

A rule for transforming a value.

PrivacyMetric

Privacy metric to compute for reidentification risk analysis.

ProfileStatus

QuasiId

A column with a semantic tag attached.

QuoteInfo

Message for infoType-dependent details parsed from quote.

Range

Generic half-open interval [start, end)

RecordCondition

A condition for determining whether a transformation should be applied to a field.

RecordKey

Message for a unique key indicating a record that contains a finding.

RecordLocation

Location of a finding within a row or record.

RecordSuppression

Configuration to suppress records whose suppression conditions evaluate to true.

RecordTransformation

RecordTransformations

A type of transformation that is applied over structured data such as a table.

RedactConfig

Redact a given value. For example, if used with an InfoTypeTransformation transforming PHONE_NUMBER, and input ‘My phone number is 206-555-0123’, the output would be ’My phone number is ’.

RedactImageRequest

Request to search for potentially sensitive info in an image and redact it by covering it with a colored rectangle.

RedactImageResponse

Results of redacting an image.

ReidentifyContentRequest

Request to re-identify an item.

ReidentifyContentResponse

Results of re-identifying an item.

ReplaceDictionaryConfig

Replace each input value with a value randomly selected from the dictionary.

ReplaceValueConfig

Replace each input value with a given Value.

ReplaceWithInfoTypeConfig

Replace each matching finding with the name of the info_type.

RiskAnalysisJobConfig

Configuration for a risk analysis job. See https://cloud.google.com/dlp/docs/concepts-risk-analysis to learn more.

Schedule

Schedule for inspect job triggers.

SensitivityScore

Score is calculated from of all elements in the data profile. A higher level means the data is more sensitive.

StatisticalTable

An auxiliary table containing statistical information on the relative frequency of different quasi-identifiers values. It has one or several quasi-identifiers columns, and one column that indicates the relative frequency of each quasi-identifier tuple. If a tuple is present in the data but not in the auxiliary table, the corresponding relative frequency is assumed to be zero (and thus, the tuple is highly reidentifiable).

StorageConfig

Shared message indicating Cloud storage type.

StorageMetadataLabel

Storage metadata label to indicate which metadata entry contains findings.

StoredInfoType

StoredInfoType resource message that contains information about the current version and any pending updates.

StoredInfoTypeConfig

Configuration for stored infoTypes. All fields and subfield are provided by the user. For more information, see https://cloud.google.com/dlp/docs/creating-custom-infotypes.

StoredInfoTypeStats

Statistics for a StoredInfoType.

StoredInfoTypeVersion

Version of a StoredInfoType, including the configuration used to build it, create timestamp, and current state.

StoredType

A reference to a StoredInfoType to use with scanning.

Table

Structured content to inspect. Up to 50,000 Values per request allowed. See https://cloud.google.com/dlp/docs/inspecting-structured-text#inspecting_a_table to learn more.

TableDataProfile

The profile for a scanned table.

TableLocation

Location of a finding within a table.

TableOptions

Instructions regarding the table content being inspected.

TimePartConfig

For use with Date, Timestamp, and TimeOfDay, extract or preserve a portion of the value.

TransformationConfig

User specified templates and configs for how to deidentify structured, unstructures, and image files. User must provide either a unstructured deidentify template or at least one redact image config.

TransformationDescription

A flattened description of a PrimitiveTransformation or RecordSuppression.

TransformationDetails

Details about a single transformation. This object contains a description of the transformation, information about whether the transformation was successfully applied, and the precise location where the transformation occurred. These details are stored in a user-specified BigQuery table.

TransformationDetailsStorageConfig

Config for storing transformation details.

TransformationErrorHandling

How to handle transformation errors during de-identification. A transformation error occurs when the requested transformation is incompatible with the data. For example, trying to de-identify an IP address using a DateShift transformation would result in a transformation error, since date info cannot be extracted from an IP address. Information about any incompatible transformations, and how they were handled, is returned in the response as part of the TransformationOverviews.

TransformationLocation

Specifies the location of a transformation.

TransformationOverview

Overview of the modifications that occurred.

TransformationResultStatus

TransformationSummary

Summary of a single transformation. Only one of ‘transformation’, ‘field_transformation’, or ‘record_suppress’ will be set.

TransientCryptoKey

Use this to have a random data crypto key generated. It will be discarded after the request finishes.

UnwrappedCryptoKey

Using raw keys is prone to security risks due to accidentally leaking the key. Choose another type of key if possible.

UpdateDeidentifyTemplateRequest

Request message for UpdateDeidentifyTemplate.

UpdateDiscoveryConfigRequest

Request message for UpdateDiscoveryConfig.

UpdateInspectTemplateRequest

Request message for UpdateInspectTemplate.

UpdateJobTriggerRequest

Request message for UpdateJobTrigger.

UpdateStoredInfoTypeRequest

Request message for UpdateStoredInfoType.

Value

Set of primitive values supported by the system. Note that for the purposes of inspection or transformation, the number of bytes considered to comprise a ‘Value’ is based on its representation as a UTF-8 encoded string. For example, if ‘integer_value’ is set to 123456789, the number of bytes would be counted as 9, even though an int64 only holds up to 8 bytes of data.

ValueFrequency

A value of a field, including its frequency.

VersionDescription

Details about each available version for an infotype.

Enums

BigQuerySchemaModification

Attributes evaluated to determine if a schema has been modified. New values may be added at a later time.

BigQueryTableModification

Attributes evaluated to determine if a table has been modified. New values may be added at a later time.

BigQueryTableType

Over time new types may be added. Currently VIEW, MATERIALIZED_VIEW, SNAPSHOT, and non-BigLake external tables are not supported.

BigQueryTableTypeCollection

Over time new types may be added. Currently VIEW, MATERIALIZED_VIEW, and SNAPSHOT are not supported.

ContentOption

Deprecated and unused.

DataProfileUpdateFrequency

How frequently data profiles can be updated. New options can be added at a later time.

DlpJobType

An enum to represent the various types of DLP jobs.

EncryptionStatus

How a resource is encrypted.

FileType

Definitions of file type groups to scan. New types will be added to this list.

InfoTypeSupportedBy

Parts of the APIs which use certain infoTypes.

Likelihood

Coarse-grained confidence level of how well a particular finding satisfies the criteria to match a particular infoType.

MatchingType

Type of the match which can be applied to different ways of matching, like Dictionary, regular expression and intersecting with findings of another info type.

MetadataType

Type of metadata containing the finding.

NullPercentageLevel

Bucketized nullness percentage levels. A higher level means a higher percentage of the column is null.

RelationalOperator

Operators available for comparing the value of fields.

ResourceVisibility

How broadly a resource has been shared. New items may be added over time. A higher number means more restricted.

StoredInfoTypeState

State of a StoredInfoType version.

TransformationContainerType

Describes functionality of a given container in its original format.

TransformationResultStatusType

Enum of possible outcomes of transformations. SUCCESS if transformation and storing of transformation was successful, otherwise, reason for not transforming.

TransformationType

An enum of rules that can be used to transform a value. Can be a record suppression, or one of the transformation rules specified under PrimitiveTransformation.

UniquenessScoreLevel

Bucketized uniqueness score levels. A higher uniqueness score is a strong signal that the column may contain a unique identifier like user id. A low value indicates that the column contains few unique values like booleans or other classifiers.