Struct google_api_proto::google::privacy::dlp::v2::dlp_service_client::DlpServiceClient

pub struct DlpServiceClient<T> { /* private fields */ }

The Cloud Data Loss Prevention (DLP) API is a service that allows clients to detect the presence of Personally Identifiable Information (PII) and other privacy-sensitive data in user-supplied, unstructured data streams, like text blocks or images. The service also includes methods for sensitive data redaction and scheduling of data scans on Google Cloud Platform based data sets.

To learn more about concepts and find how-to guides see https://cloud.google.com/dlp/docs/.

Implementations

impl<T> DlpServiceClient<T>where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F ) -> DlpServiceClient<InterceptedService<T, F>>where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn inspect_content( &mut self, request: impl IntoRequest<InspectContentRequest> ) -> Result<Response<InspectContentResponse>, Status>

Finds potentially sensitive info in content. This method has limits on input size, processing time, and output size.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

For how to guides, see https://cloud.google.com/dlp/docs/inspecting-images and https://cloud.google.com/dlp/docs/inspecting-text,

pub async fn redact_image( &mut self, request: impl IntoRequest<RedactImageRequest> ) -> Result<Response<RedactImageResponse>, Status>

Redacts potentially sensitive info from an image. This method has limits on input size, processing time, and output size. See https://cloud.google.com/dlp/docs/redacting-sensitive-data-images to learn more.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

pub async fn deidentify_content( &mut self, request: impl IntoRequest<DeidentifyContentRequest> ) -> Result<Response<DeidentifyContentResponse>, Status>

De-identifies potentially sensitive info from a ContentItem. This method has limits on input size and output size. See https://cloud.google.com/dlp/docs/deidentify-sensitive-data to learn more.

When no InfoTypes or CustomInfoTypes are specified in this request, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

pub async fn reidentify_content( &mut self, request: impl IntoRequest<ReidentifyContentRequest> ) -> Result<Response<ReidentifyContentResponse>, Status>

Re-identifies content that has been de-identified. See https://cloud.google.com/dlp/docs/pseudonymization#re-identification_in_free_text_code_example to learn more.

pub async fn list_info_types( &mut self, request: impl IntoRequest<ListInfoTypesRequest> ) -> Result<Response<ListInfoTypesResponse>, Status>

Returns a list of the sensitive information types that DLP API supports. See https://cloud.google.com/dlp/docs/infotypes-reference to learn more.

pub async fn create_inspect_template( &mut self, request: impl IntoRequest<CreateInspectTemplateRequest> ) -> Result<Response<InspectTemplate>, Status>

Creates an InspectTemplate for reusing frequently used configuration for inspecting content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

pub async fn update_inspect_template( &mut self, request: impl IntoRequest<UpdateInspectTemplateRequest> ) -> Result<Response<InspectTemplate>, Status>

Updates the InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

pub async fn get_inspect_template( &mut self, request: impl IntoRequest<GetInspectTemplateRequest> ) -> Result<Response<InspectTemplate>, Status>

Gets an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

pub async fn list_inspect_templates( &mut self, request: impl IntoRequest<ListInspectTemplatesRequest> ) -> Result<Response<ListInspectTemplatesResponse>, Status>

Lists InspectTemplates. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

pub async fn delete_inspect_template( &mut self, request: impl IntoRequest<DeleteInspectTemplateRequest> ) -> Result<Response<()>, Status>

Deletes an InspectTemplate. See https://cloud.google.com/dlp/docs/creating-templates to learn more.

pub async fn create_deidentify_template( &mut self, request: impl IntoRequest<CreateDeidentifyTemplateRequest> ) -> Result<Response<DeidentifyTemplate>, Status>

Creates a DeidentifyTemplate for reusing frequently used configuration for de-identifying content, images, and storage. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

pub async fn update_deidentify_template( &mut self, request: impl IntoRequest<UpdateDeidentifyTemplateRequest> ) -> Result<Response<DeidentifyTemplate>, Status>

Updates the DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

pub async fn get_deidentify_template( &mut self, request: impl IntoRequest<GetDeidentifyTemplateRequest> ) -> Result<Response<DeidentifyTemplate>, Status>

Gets a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

pub async fn list_deidentify_templates( &mut self, request: impl IntoRequest<ListDeidentifyTemplatesRequest> ) -> Result<Response<ListDeidentifyTemplatesResponse>, Status>

Lists DeidentifyTemplates. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

pub async fn delete_deidentify_template( &mut self, request: impl IntoRequest<DeleteDeidentifyTemplateRequest> ) -> Result<Response<()>, Status>

Deletes a DeidentifyTemplate. See https://cloud.google.com/dlp/docs/creating-templates-deid to learn more.

pub async fn create_job_trigger( &mut self, request: impl IntoRequest<CreateJobTriggerRequest> ) -> Result<Response<JobTrigger>, Status>

Creates a job trigger to run DLP actions such as scanning storage for sensitive information on a set schedule. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

pub async fn update_job_trigger( &mut self, request: impl IntoRequest<UpdateJobTriggerRequest> ) -> Result<Response<JobTrigger>, Status>

Updates a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

pub async fn hybrid_inspect_job_trigger( &mut self, request: impl IntoRequest<HybridInspectJobTriggerRequest> ) -> Result<Response<HybridInspectResponse>, Status>

Inspect hybrid content and store findings to a trigger. The inspection will be processed asynchronously. To review the findings monitor the jobs within the trigger.

pub async fn get_job_trigger( &mut self, request: impl IntoRequest<GetJobTriggerRequest> ) -> Result<Response<JobTrigger>, Status>

Gets a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

pub async fn list_job_triggers( &mut self, request: impl IntoRequest<ListJobTriggersRequest> ) -> Result<Response<ListJobTriggersResponse>, Status>

Lists job triggers. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

pub async fn delete_job_trigger( &mut self, request: impl IntoRequest<DeleteJobTriggerRequest> ) -> Result<Response<()>, Status>

Deletes a job trigger. See https://cloud.google.com/dlp/docs/creating-job-triggers to learn more.

pub async fn activate_job_trigger( &mut self, request: impl IntoRequest<ActivateJobTriggerRequest> ) -> Result<Response<DlpJob>, Status>

Activate a job trigger. Causes the immediate execute of a trigger instead of waiting on the trigger event to occur.

pub async fn create_discovery_config( &mut self, request: impl IntoRequest<CreateDiscoveryConfigRequest> ) -> Result<Response<DiscoveryConfig>, Status>

Creates a config for discovery to scan and profile storage.

pub async fn update_discovery_config( &mut self, request: impl IntoRequest<UpdateDiscoveryConfigRequest> ) -> Result<Response<DiscoveryConfig>, Status>

Updates a discovery configuration.

pub async fn get_discovery_config( &mut self, request: impl IntoRequest<GetDiscoveryConfigRequest> ) -> Result<Response<DiscoveryConfig>, Status>

Gets a discovery configuration.

pub async fn list_discovery_configs( &mut self, request: impl IntoRequest<ListDiscoveryConfigsRequest> ) -> Result<Response<ListDiscoveryConfigsResponse>, Status>

Lists discovery configurations.

pub async fn delete_discovery_config( &mut self, request: impl IntoRequest<DeleteDiscoveryConfigRequest> ) -> Result<Response<()>, Status>

Deletes a discovery configuration.

pub async fn create_dlp_job( &mut self, request: impl IntoRequest<CreateDlpJobRequest> ) -> Result<Response<DlpJob>, Status>

Creates a new job to inspect storage or calculate risk metrics. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

When no InfoTypes or CustomInfoTypes are specified in inspect jobs, the system will automatically choose what detectors to run. By default this may be all types, but may change over time as detectors are updated.

pub async fn list_dlp_jobs( &mut self, request: impl IntoRequest<ListDlpJobsRequest> ) -> Result<Response<ListDlpJobsResponse>, Status>

Lists DlpJobs that match the specified filter in the request. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

pub async fn get_dlp_job( &mut self, request: impl IntoRequest<GetDlpJobRequest> ) -> Result<Response<DlpJob>, Status>

Gets the latest state of a long-running DlpJob. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

pub async fn delete_dlp_job( &mut self, request: impl IntoRequest<DeleteDlpJobRequest> ) -> Result<Response<()>, Status>

Deletes a long-running DlpJob. This method indicates that the client is no longer interested in the DlpJob result. The job will be canceled if possible. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

pub async fn cancel_dlp_job( &mut self, request: impl IntoRequest<CancelDlpJobRequest> ) -> Result<Response<()>, Status>

Starts asynchronous cancellation on a long-running DlpJob. The server makes a best effort to cancel the DlpJob, but success is not guaranteed. See https://cloud.google.com/dlp/docs/inspecting-storage and https://cloud.google.com/dlp/docs/compute-risk-analysis to learn more.

pub async fn create_stored_info_type( &mut self, request: impl IntoRequest<CreateStoredInfoTypeRequest> ) -> Result<Response<StoredInfoType>, Status>

Creates a pre-built stored infoType to be used for inspection. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

pub async fn update_stored_info_type( &mut self, request: impl IntoRequest<UpdateStoredInfoTypeRequest> ) -> Result<Response<StoredInfoType>, Status>

Updates the stored infoType by creating a new version. The existing version will continue to be used until the new version is ready. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

pub async fn get_stored_info_type( &mut self, request: impl IntoRequest<GetStoredInfoTypeRequest> ) -> Result<Response<StoredInfoType>, Status>

Gets a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

pub async fn list_stored_info_types( &mut self, request: impl IntoRequest<ListStoredInfoTypesRequest> ) -> Result<Response<ListStoredInfoTypesResponse>, Status>

Lists stored infoTypes. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

pub async fn delete_stored_info_type( &mut self, request: impl IntoRequest<DeleteStoredInfoTypeRequest> ) -> Result<Response<()>, Status>

Deletes a stored infoType. See https://cloud.google.com/dlp/docs/creating-stored-infotypes to learn more.

pub async fn hybrid_inspect_dlp_job( &mut self, request: impl IntoRequest<HybridInspectDlpJobRequest> ) -> Result<Response<HybridInspectResponse>, Status>

Inspect hybrid content and store findings to a job. To review the findings, inspect the job. Inspection will occur asynchronously.

pub async fn finish_dlp_job( &mut self, request: impl IntoRequest<FinishDlpJobRequest> ) -> Result<Response<()>, Status>

Finish a running hybrid DlpJob. Triggers the finalization steps and running of any enabled actions that have not yet run.