Module google_api_proto::google::rpc

Modules

bad_request

Nested message and enum types in BadRequest.

context

help

Nested message and enum types in Help.

precondition_failure

Nested message and enum types in PreconditionFailure.

quota_failure

Nested message and enum types in QuotaFailure.

Structs

BadRequest

Describes violations in a client request. This error type focuses on the syntactic aspects of the request.

DebugInfo

Describes additional debugging info.

ErrorInfo

Describes the cause of the error with structured details.

Help

Provides links to documentation or for performing an out of band action.

HttpHeader

Represents an HTTP header.

HttpRequest

Represents an HTTP request.

HttpResponse

Represents an HTTP response.

LocalizedMessage

Provides a localized error message that is safe to return to the user which can be attached to an RPC error.

PreconditionFailure

Describes what preconditions have failed.

QuotaFailure

Describes how a quota check failed.

RequestInfo

Contains metadata about the request that clients can attach when filing a bug or providing other forms of feedback.

ResourceInfo

Describes the resource that is being accessed.

RetryInfo

Describes when the clients can retry a failed request. Clients could ignore the recommendation here or retry when this information is missing from error responses.

Status

The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

Enums

Code

The canonical error codes for gRPC APIs.