Module google_api_proto::google::shopping::css::v1

Modules

account

Nested message and enum types in Account.

account_label

Nested message and enum types in AccountLabel.

account_labels_service_client

Generated client implementations.

accounts_service_client

Generated client implementations.

css_product_inputs_service_client

Generated client implementations.

css_product_status

Nested message and enum types in CssProductStatus.

css_products_service_client

Generated client implementations.

Structs

Account

Information about CSS/MC account.

AccountLabel

Label assigned by CSS domain or CSS group to one of its sub-accounts.

Attributes

Attributes for CSS Product.

Certification

The certification for the product.

CreateAccountLabelRequest

Request message for the ‘CreateAccountLanel’ method.

CssProduct

The processed CSS Product(a.k.a Aggregate Offer internally).

CssProductInput

This resource represents input data you submit for a CSS Product, not the processed CSS Product that you see in CSS Center, in Shopping Ads, or across Google surfaces.

CssProductStatus

The status of the Css Product, data validation issues, that is, information about the Css Product computed asynchronously.

DeleteAccountLabelRequest

Request message for the ‘DeleteAccountLabel’ method.

DeleteCssProductInputRequest

Request message for the DeleteCssProductInput method.

GetAccountRequest

The request message for the GetAccount method.

GetCssProductRequest

The request message for the GetCssProduct method.

InsertCssProductInputRequest

Request message for the InsertCssProductInput method.

ListAccountLabelsRequest

Request message for the ListAccountLabels method.

ListAccountLabelsResponse

Response message for the ListAccountLabels method.

ListChildAccountsRequest

The request message for the ListChildAccounts method.

ListChildAccountsResponse

Response message for the ListChildAccounts method.

ListCssProductsRequest

Request message for the ListCssProducts method.

ListCssProductsResponse

Response message for the ListCssProducts method.

ProductDetail

The product details.

ProductDimension

The dimension of the product.

ProductWeight

The weight of the product.

UpdateAccountLabelRequest

Request message for the UpdateAccountLabel method.

UpdateAccountLabelsRequest

The request message for the UpdateLabels method.