Module google_api_proto::google::shopping::merchant::quota::v1beta

Modules

quota_service_client

Generated client implementations.

Structs

ListQuotaGroupsRequest

Request message for the ListQuotaGroups method.

ListQuotaGroupsResponse

Response message for the ListMethodGroups method.

MethodDetails

The method details per method in the Merchant API.

QuotaGroup

The group information for methods in the Merchant API. The quota is shared between all methods in the group. Even if none of the methods within the group have usage the information for the group is returned.