Module google_api_proto::google::shopping::merchant::reports::v1beta

Modules

best_sellers_product_cluster_view

Nested message and enum types in BestSellersProductClusterView.

marketing_method

Nested message and enum types in MarketingMethod.

product_view

Nested message and enum types in ProductView.

relative_demand

Nested message and enum types in RelativeDemand.

relative_demand_change_type

Nested message and enum types in RelativeDemandChangeType.

report_granularity

Nested message and enum types in ReportGranularity.

report_service_client

Generated client implementations.

traffic_source

Nested message and enum types in TrafficSource.

Structs

BestSellersBrandView

Fields available for query in best_sellers_brand_view table.

BestSellersProductClusterView

Fields available for query in best_sellers_product_cluster_view table.

CompetitiveVisibilityBenchmarkView

Fields available for query in competitive_visibility_benchmark_view table.

CompetitiveVisibilityCompetitorView

Fields available for query in competitive_visibility_competitor_view table.

CompetitiveVisibilityTopMerchantView

Fields available for query in competitive_visibility_top_merchant_view table.

MarketingMethod

Marketing method used to promote your products on Google (organic versus ads).

PriceCompetitivenessProductView

Fields available for query in price_competitiveness_product_view table.

PriceInsightsProductView

Fields available for query in price_insights_product_view table.

ProductPerformanceView

Fields available for query in product_performance_view table.

ProductView

Fields available for query in product_view table.

RelativeDemand

Relative demand of a product cluster or brand in the Best sellers report.

RelativeDemandChangeType

Relative demand of a product cluster or brand in the Best sellers report compared to the previous time period.

ReportGranularity

Granularity of the Best sellers report. Best sellers reports are computed over a week and a month timeframe.

ReportRow

Result row returned from the search query.

SearchRequest

Request message for the ReportService.Search method.

SearchResponse

Response message for the ReportService.Search method.

TrafficSource

Traffic source of impressions in the Competitive visibility report.