Module google_api_proto::google::spanner::admin::instance::v1

Modules

instance

Nested message and enum types in Instance.

instance_admin_client

Generated client implementations.

instance_config

Nested message and enum types in InstanceConfig.

replica_info

Nested message and enum types in ReplicaInfo.

Structs

CreateInstanceConfigMetadata

Metadata type for the operation returned by [CreateInstanceConfig][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstanceConfig].

CreateInstanceConfigRequest

The request for [CreateInstanceConfigRequest][InstanceAdmin.CreateInstanceConfigRequest].

CreateInstanceMetadata

Metadata type for the operation returned by [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

CreateInstanceRequest

The request for [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

DeleteInstanceConfigRequest

The request for [DeleteInstanceConfigRequest][InstanceAdmin.DeleteInstanceConfigRequest].

DeleteInstanceRequest

The request for [DeleteInstance][google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstance].

GetInstanceConfigRequest

The request for [GetInstanceConfigRequest][google.spanner.admin.instance.v1.InstanceAdmin.GetInstanceConfig].

GetInstanceRequest

The request for [GetInstance][google.spanner.admin.instance.v1.InstanceAdmin.GetInstance].

Instance

An isolated set of Cloud Spanner resources on which databases can be hosted.

InstanceConfig

A possible configuration for a Cloud Spanner instance. Configurations define the geographic placement of nodes and their replication.

ListInstanceConfigOperationsRequest

The request for [ListInstanceConfigOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigOperations].

ListInstanceConfigOperationsResponse

The response for [ListInstanceConfigOperations][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigOperations].

ListInstanceConfigsRequest

The request for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstanceConfigsResponse

The response for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstancesRequest

The request for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

ListInstancesResponse

The response for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

OperationProgress

Encapsulates progress related information for a Cloud Spanner long running instance operations.

ReplicaInfo

UpdateInstanceConfigMetadata

Metadata type for the operation returned by [UpdateInstanceConfig][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstanceConfig].

UpdateInstanceConfigRequest

The request for [UpdateInstanceConfigRequest][InstanceAdmin.UpdateInstanceConfigRequest].

UpdateInstanceMetadata

Metadata type for the operation returned by [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].

UpdateInstanceRequest

The request for [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].