Module google_api_proto::google::spanner::executor::v1

Modules

admin_action

Nested message and enum types in AdminAction.

change_stream_record

Nested message and enum types in ChangeStreamRecord.

child_partitions_record

Nested message and enum types in ChildPartitionsRecord.

concurrency

Nested message and enum types in Concurrency.

data_change_record

Nested message and enum types in DataChangeRecord.

finish_transaction_action

Nested message and enum types in FinishTransactionAction.

key_range

Nested message and enum types in KeyRange.

mutation_action

Nested message and enum types in MutationAction.

partitioned_update_action

Nested message and enum types in PartitionedUpdateAction.

query_action

Nested message and enum types in QueryAction.

spanner_action

Nested message and enum types in SpannerAction.

spanner_executor_proxy_client

Generated client implementations.

start_batch_transaction_action

Nested message and enum types in StartBatchTransactionAction.

value

Nested message and enum types in Value.

Structs

AdminAction

AdminAction defines all the cloud spanner admin actions, including instance/database admin ops, backup ops and operation actions.

AdminResult

AdminResult contains admin action results, for database/backup/operation.

BatchDmlAction

Batch of DML statements invoked using batched execution.

BatchPartition

Identifies a database partition generated for a particular read or query. To read rows from the partition, use ExecutePartitionAction.

CancelOperationAction

Action that cancels an operation.

ChangeStreamRecord

Raw ChangeStream records. Encodes one of: DataChangeRecord, HeartbeatRecord, ChildPartitionsRecord returned from the ChangeStream API.

ChildPartitionsRecord

ChangeStream child partition record.

CloseBatchTransactionAction

Closes or cleans up the currently opened batch read-only transaction.

CloudBackupResponse

CloudBackupResponse contains results returned by cloud backup related actions.

CloudDatabaseResponse

CloudDatabaseResponse contains results returned by cloud database related actions.

CloudInstanceConfigResponse

CloudInstanceConfigResponse contains results returned by cloud instance config related actions.

CloudInstanceResponse

CloudInstanceResponse contains results returned by cloud instance related actions.

ColumnMetadata

ColumnMetadata represents metadata of a single column.

Concurrency

Concurrency for read-only transactions.

CopyCloudBackupAction

Action that copies a Cloud Spanner database backup.

CreateCloudBackupAction

Action that creates a Cloud Spanner database backup.

CreateCloudDatabaseAction

Action that creates a Cloud Spanner database.

CreateCloudInstanceAction

Action that creates a Cloud Spanner instance.

CreateUserInstanceConfigAction

Action that creates a user instance config.

DataChangeRecord

ChangeStream data change record.

DeleteCloudBackupAction

Action that deletes a Cloud Spanner database backup.

DeleteCloudInstanceAction

Action that deletes a Cloud Spanner instance.

DeleteUserInstanceConfigAction

Action that deletes a user instance configs.

DmlAction

A single DML statement.

DropCloudDatabaseAction

Action that drops a Cloud Spanner database.

ExecuteChangeStreamQuery

Execute a change stream TVF query.

ExecutePartitionAction

Performs a read or query for the given partitions. This action must be executed in the context of the same transaction that was used to generate given partitions.

FinishTransactionAction

FinishTransactionAction defines an action of finishing a transaction.

GenerateDbPartitionsForQueryAction

Generate database partitions for the given query. Successful outcomes will contain database partitions in the db_partition field.

GenerateDbPartitionsForReadAction

Generate database partitions for the given read. Successful outcomes will contain database partitions in the db_partition field.

GetCloudBackupAction

Action that gets a Cloud Spanner database backup.

GetCloudDatabaseAction

Action that gets a Cloud Spanner database.

GetCloudInstanceAction

Action that retrieves a Cloud Spanner instance.

GetCloudInstanceConfigAction

Action that gets a user instance config.

GetOperationAction

Action that gets an operation.

HeartbeatRecord

ChangeStream heartbeat record.

KeyRange

KeyRange represents a range of rows in a table or index.

KeySet

KeySet defines a collection of Spanner keys and/or key ranges. All the keys are expected to be in the same table. The keys need not be sorted in any particular way.

ListCloudBackupOperationsAction

Action that lists Cloud Spanner database backup operations.

ListCloudBackupsAction

Action that lists Cloud Spanner database backups.

ListCloudDatabaseOperationsAction

Action that lists Cloud Spanner database operations.

ListCloudDatabasesAction

Action that lists Cloud Spanner databases.

ListCloudInstanceConfigsAction

Action that lists user instance configs.

ListCloudInstancesAction

Action that lists Cloud Spanner databases.

MutationAction

A single mutation request.

OperationResponse

OperationResponse contains results returned by operation related actions.

PartitionedUpdateAction

PartitionedUpdateAction defines an action to execute a partitioned DML which runs different partitions in parallel.

QueryAction

A SQL query request.

QueryResult

QueryResult contains result of a Query.

ReadAction

A single read request.

ReadResult

ReadResult contains rows read.

ReconfigureCloudDatabaseAction

Action that reconfigures a Cloud Spanner database.

RestoreCloudDatabaseAction

Action that restores a Cloud Spanner database from a backup.

SpannerAction

SpannerAction defines a primitive action that can be performed against Spanner, such as begin or commit a transaction, or perform a read or mutation.

SpannerActionOutcome

SpannerActionOutcome defines a result of execution of a single SpannerAction.

SpannerAsyncActionRequest

Request to executor service that start a new Spanner action.

SpannerAsyncActionResponse

Response from executor service.

StartBatchTransactionAction

Starts a batch read-only transaction in executor. Successful outcomes of this action will contain batch_txn_id–the identificator that can be used to start the same transaction in other Executors to parallelize partition processing.

StartTransactionAction

StartTransactionAction defines an action of initializing a transaction.

TableMetadata

TableMetadata contains metadata of a single table.

TransactionExecutionOptions

Options for executing the transaction.

UpdateCloudBackupAction

Action that updates a Cloud Spanner database backup.

UpdateCloudDatabaseAction

Action that updates a Cloud Spanner database.

UpdateCloudDatabaseDdlAction

Action that updates the schema of a Cloud Spanner database.

UpdateCloudInstanceAction

Action that updates a Cloud Spanner instance.

UpdateUserInstanceConfigAction

Action that updates a user instance config.

Value

Value represents a single value that can be read or written to/from Spanner.

ValueList

List of values.

WriteMutationsAction

WriteMutationAction defines an action of flushing the mutation so they are visible to subsequent operations in the transaction.