Module google_api_proto::google::storage::v2

Modules

bucket

Nested message and enum types in Bucket.

compose_object_request

Nested message and enum types in ComposeObjectRequest.

object

Nested message and enum types in Object.

query_write_status_response

Nested message and enum types in QueryWriteStatusResponse.

service_constants

Nested message and enum types in ServiceConstants.

storage_client

Generated client implementations.

write_object_request

Nested message and enum types in WriteObjectRequest.

write_object_response

Nested message and enum types in WriteObjectResponse.

Structs

Bucket

A bucket.

BucketAccessControl

An access-control entry.

ChecksummedData

Message used to convey content being read or written, along with an optional checksum.

CommonObjectRequestParams

Parameters that can be passed to any object request.

CommonRequestParams

Parameters that can be passed to any request.

ComposeObjectRequest

Request message for ComposeObject.

ContentRange

Specifies a requested range of bytes to download.

CreateBucketRequest

Request message for CreateBucket.

CreateHmacKeyRequest

Request message for CreateHmacKey.

CreateHmacKeyResponse

Create hmac response. The only time the secret for an HMAC will be returned.

CreateNotificationRequest

Request message for CreateNotification.

DeleteBucketRequest

Request message for DeleteBucket.

DeleteHmacKeyRequest

Request object to delete a given HMAC key.

DeleteNotificationRequest

Request message for DeleteNotification.

DeleteObjectRequest

Message for deleting an object. Either bucket and object or upload_id must be set (but not both).

GetBucketRequest

Request message for GetBucket.

GetHmacKeyRequest

Request object to get metadata on a given HMAC key.

GetNotificationRequest

Request message for GetNotification.

GetObjectRequest

Request message for GetObject.

GetServiceAccountRequest

Request message for GetServiceAccount.

HmacKeyMetadata

Hmac Key Metadata, which includes all information other than the secret.

ListBucketsRequest

Request message for ListBuckets.

ListBucketsResponse

The result of a call to Buckets.ListBuckets

ListHmacKeysRequest

Request to fetch a list of HMAC keys under a given project.

ListHmacKeysResponse

Hmac key list response with next page information.

ListNotificationsRequest

Request message for ListNotifications.

ListNotificationsResponse

The result of a call to Notifications.ListNotifications

ListObjectsRequest

Request message for ListObjects.

ListObjectsResponse

The result of a call to Objects.ListObjects

LockBucketRetentionPolicyRequest

Request message for LockBucketRetentionPolicyRequest.

Notification

A directive to publish Pub/Sub notifications upon changes to a bucket.

Object

An object.

ObjectAccessControl

An access-control entry.

ObjectChecksums

Message used for storing full (not subrange) object checksums.

Owner

The owner of a specific resource.

ProjectTeam

Represents the Viewers, Editors, or Owners of a given project.

QueryWriteStatusRequest

Request object for QueryWriteStatus.

QueryWriteStatusResponse

Response object for QueryWriteStatus.

ReadObjectRequest

Request message for ReadObject.

ReadObjectResponse

Response message for GetObject.

RewriteObjectRequest

Request message for RewriteObject.

RewriteResponse

A rewrite response.

ServiceAccount

A service account, owned by Cloud Storage, which may be used when taking action on behalf of a given project, for example to publish Pub/Sub notifications or to retrieve security keys.

ServiceConstants

Shared constants.

StartResumableWriteRequest

Request message StartResumableWrite.

StartResumableWriteResponse

Response object for StartResumableWrite.

UpdateBucketRequest

Request for UpdateBucket method.

UpdateHmacKeyRequest

Request object to update an HMAC key state.

UpdateObjectRequest

Request message for UpdateObject.

WriteObjectRequest

Request message for WriteObject.

WriteObjectResponse

Response message for WriteObject.

WriteObjectSpec

Describes an attempt to insert an object, possibly over multiple requests.

Enums

PredefinedBucketAcl

A set of predefined, or “canned,” ACLs that can be set on a bucket instead of manually specifying a complete ACL.

PredefinedObjectAcl

Predefined or “canned” aliases for sets of specific object ACL entries.