Struct google_api_proto::google::storage::v2::storage_client::StorageClient

pub struct StorageClient<T> { /* fields omitted */ }

API Overview and Naming Syntax

The Cloud Storage gRPC API allows applications to read and write data through the abstractions of buckets and objects. For a description of these abstractions please see https://cloud.google.com/storage/docs.

Resources are named as follows:

Projects are referred to as they are defined by the Resource Manager API, using strings like projects/123456 or projects/my-string-id.

Buckets are named using string names of the form: projects/{project}/buckets/{bucket} For globally unique buckets, _ may be substituted for the project.

Objects are uniquely identified by their name along with the name of the bucket they belong to, as separate strings in this API. For example:

ReadObjectRequest { bucket: ‘projects/_/buckets/my-bucket’ object: ‘my-object’ } Note that object names can contain / characters, which are treated as any other character (no special directory semantics).

Implementations

impl<T> StorageClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> StorageClient<InterceptedService<T, F>> where
F: Interceptor,
T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>,
<T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_gzip(self) -> Self

Compress requests with gzip.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_gzip(self) -> Self

Enable decompressing responses with gzip.

pub async fn delete_bucket(
&mut self,
request: impl IntoRequest<DeleteBucketRequest>
) -> Result<Response<()>, Status>

Permanently deletes an empty bucket.

pub async fn get_bucket(
&mut self,
request: impl IntoRequest<GetBucketRequest>
) -> Result<Response<Bucket>, Status>

Returns metadata for the specified bucket.

pub async fn create_bucket(
&mut self,
request: impl IntoRequest<CreateBucketRequest>
) -> Result<Response<Bucket>, Status>

Creates a new bucket.

pub async fn list_buckets(
&mut self,
request: impl IntoRequest<ListBucketsRequest>
) -> Result<Response<ListBucketsResponse>, Status>

Retrieves a list of buckets for a given project.

pub async fn lock_bucket_retention_policy(
&mut self,
request: impl IntoRequest<LockBucketRetentionPolicyRequest>
) -> Result<Response<Bucket>, Status>

Locks retention policy on a bucket.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the IAM policy for a specified bucket.

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Updates an IAM policy for the specified bucket.

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Tests a set of permissions on the given bucket to see which, if any, are held by the caller.

pub async fn update_bucket(
&mut self,
request: impl IntoRequest<UpdateBucketRequest>
) -> Result<Response<Bucket>, Status>

Updates a bucket. Equivalent to JSON API’s storage.buckets.patch method.

pub async fn delete_notification(
&mut self,
request: impl IntoRequest<DeleteNotificationRequest>
) -> Result<Response<()>, Status>

Permanently deletes a notification subscription.

pub async fn get_notification(
&mut self,
request: impl IntoRequest<GetNotificationRequest>
) -> Result<Response<Notification>, Status>

View a notification config.

pub async fn create_notification(
&mut self,
request: impl IntoRequest<CreateNotificationRequest>
) -> Result<Response<Notification>, Status>

Creates a notification subscription for a given bucket. These notifications, when triggered, publish messages to the specified Pub/Sub topics. See https://cloud.google.com/storage/docs/pubsub-notifications.

pub async fn list_notifications(
&mut self,
request: impl IntoRequest<ListNotificationsRequest>
) -> Result<Response<ListNotificationsResponse>, Status>

Retrieves a list of notification subscriptions for a given bucket.

pub async fn compose_object(
&mut self,
request: impl IntoRequest<ComposeObjectRequest>
) -> Result<Response<Object>, Status>

Concatenates a list of existing objects into a new object in the same bucket.

pub async fn delete_object(
&mut self,
request: impl IntoRequest<DeleteObjectRequest>
) -> Result<Response<()>, Status>

Deletes an object and its metadata. Deletions are permanent if versioning is not enabled for the bucket, or if the generation parameter is used.

pub async fn get_object(
&mut self,
request: impl IntoRequest<GetObjectRequest>
) -> Result<Response<Object>, Status>

Retrieves an object’s metadata.

pub async fn read_object(
&mut self,
request: impl IntoRequest<ReadObjectRequest>
) -> Result<Response<Streaming<ReadObjectResponse>>, Status>

Reads an object’s data.

pub async fn update_object(
&mut self,
request: impl IntoRequest<UpdateObjectRequest>
) -> Result<Response<Object>, Status>

Updates an object’s metadata. Equivalent to JSON API’s storage.objects.patch.

pub async fn write_object(
&mut self,
request: impl IntoStreamingRequest<Message = WriteObjectRequest>
) -> Result<Response<WriteObjectResponse>, Status>

Stores a new object and metadata.

An object can be written either in a single message stream or in a resumable sequence of message streams. To write using a single stream, the client should include in the first message of the stream an WriteObjectSpec describing the destination bucket, object, and any preconditions. Additionally, the final message must set ‘finish_write’ to true, or else it is an error.

For a resumable write, the client should instead call StartResumableWrite() and provide that method an WriteObjectSpec. They should then attach the returned upload_id to the first message of each following call to Create. If there is an error or the connection is broken during the resumable Create(), the client should check the status of the Create() by calling QueryWriteStatus() and continue writing from the returned persisted_size. This may be less than the amount of data the client previously sent.

The service will not view the object as complete until the client has sent a WriteObjectRequest with finish_write set to true. Sending any requests on a stream after sending a request with finish_write set to true will cause an error. The client should check the response it receives to determine how much data the service was able to commit and whether the service views the object as complete.

pub async fn list_objects(
&mut self,
request: impl IntoRequest<ListObjectsRequest>
) -> Result<Response<ListObjectsResponse>, Status>

Retrieves a list of objects matching the criteria.

pub async fn rewrite_object(
&mut self,
request: impl IntoRequest<RewriteObjectRequest>
) -> Result<Response<RewriteResponse>, Status>

Rewrites a source object to a destination object. Optionally overrides metadata.

pub async fn start_resumable_write(
&mut self,
request: impl IntoRequest<StartResumableWriteRequest>
) -> Result<Response<StartResumableWriteResponse>, Status>

Starts a resumable write. How long the write operation remains valid, and what happens when the write operation becomes invalid, are service-dependent.

pub async fn query_write_status(
&mut self,
request: impl IntoRequest<QueryWriteStatusRequest>
) -> Result<Response<QueryWriteStatusResponse>, Status>

Determines the persisted_size for an object that is being written, which can then be used as the write_offset for the next Write() call.

If the object does not exist (i.e., the object has been deleted, or the first Write() has not yet reached the service), this method returns the error NOT_FOUND.

The client may call QueryWriteStatus() at any time to determine how much data has been processed for this object. This is useful if the client is buffering data and needs to know which data can be safely evicted. For any sequence of QueryWriteStatus() calls for a given object name, the sequence of returned persisted_size values will be non-decreasing.

pub async fn get_service_account(
&mut self,
request: impl IntoRequest<GetServiceAccountRequest>
) -> Result<Response<ServiceAccount>, Status>

Retrieves the name of a project’s Google Cloud Storage service account.

pub async fn create_hmac_key(
&mut self,
request: impl IntoRequest<CreateHmacKeyRequest>
) -> Result<Response<CreateHmacKeyResponse>, Status>

Creates a new HMAC key for the given service account.

pub async fn delete_hmac_key(
&mut self,
request: impl IntoRequest<DeleteHmacKeyRequest>
) -> Result<Response<()>, Status>

Deletes a given HMAC key. Key must be in an INACTIVE state.

pub async fn get_hmac_key(
&mut self,
request: impl IntoRequest<GetHmacKeyRequest>
) -> Result<Response<HmacKeyMetadata>, Status>

Gets an existing HMAC key metadata for the given id.

pub async fn list_hmac_keys(
&mut self,
request: impl IntoRequest<ListHmacKeysRequest>
) -> Result<Response<ListHmacKeysResponse>, Status>

Lists HMAC keys under a given project with the additional filters provided.

pub async fn update_hmac_key(
&mut self,
request: impl IntoRequest<UpdateHmacKeyRequest>
) -> Result<Response<HmacKeyMetadata>, Status>

Updates a given HMAC key state between ACTIVE and INACTIVE.