Struct google_api_proto::google::storagetransfer::v1::MetadataOptions

pub struct MetadataOptions {
    pub symlink: i32,
    pub mode: i32,
    pub gid: i32,
    pub uid: i32,
    pub acl: i32,
    pub storage_class: i32,
    pub temporary_hold: i32,
    pub kms_key: i32,
    pub time_created: i32,
}

Specifies the metadata options for running a transfer.

Fields
symlink: i32

Specifies how symlinks should be handled by the transfer. By default, symlinks are not preserved. Only applicable to transfers involving POSIX file systems, and ignored for other transfers.

mode: i32

Specifies how each file’s mode attribute should be handled by the transfer. By default, mode is not preserved. Only applicable to transfers involving POSIX file systems, and ignored for other transfers.

gid: i32

Specifies how each file’s POSIX group ID (GID) attribute should be handled by the transfer. By default, GID is not preserved. Only applicable to transfers involving POSIX file systems, and ignored for other transfers.

uid: i32

Specifies how each file’s POSIX user ID (UID) attribute should be handled by the transfer. By default, UID is not preserved. Only applicable to transfers involving POSIX file systems, and ignored for other transfers.

acl: i32

Specifies how each object’s ACLs should be preserved for transfers between Google Cloud Storage buckets. If unspecified, the default behavior is the same as ACL_DESTINATION_BUCKET_DEFAULT.

storage_class: i32

Specifies the storage class to set on objects being transferred to Google Cloud Storage buckets. If unspecified, the default behavior is the same as [STORAGE_CLASS_DESTINATION_BUCKET_DEFAULT][google.storagetransfer.v1.MetadataOptions.StorageClass.STORAGE_CLASS_DESTINATION_BUCKET_DEFAULT].

temporary_hold: i32

Specifies how each object’s temporary hold status should be preserved for transfers between Google Cloud Storage buckets. If unspecified, the default behavior is the same as [TEMPORARY_HOLD_PRESERVE][google.storagetransfer.v1.MetadataOptions.TemporaryHold.TEMPORARY_HOLD_PRESERVE].

kms_key: i32

Specifies how each object’s Cloud KMS customer-managed encryption key (CMEK) is preserved for transfers between Google Cloud Storage buckets. If unspecified, the default behavior is the same as [KMS_KEY_DESTINATION_BUCKET_DEFAULT][google.storagetransfer.v1.MetadataOptions.KmsKey.KMS_KEY_DESTINATION_BUCKET_DEFAULT].

time_created: i32

Specifies how each object’s timeCreated metadata is preserved for transfers between Google Cloud Storage buckets. If unspecified, the default behavior is the same as [TIME_CREATED_SKIP][google.storagetransfer.v1.MetadataOptions.TimeCreated.TIME_CREATED_SKIP].

Implementations

impl MetadataOptions

pub fn symlink(&self) -> Symlink

Returns the enum value of symlink, or the default if the field is set to an invalid enum value.

pub fn set_symlink(&mut self, value: Symlink)

Sets symlink to the provided enum value.

pub fn mode(&self) -> Mode

Returns the enum value of mode, or the default if the field is set to an invalid enum value.

pub fn set_mode(&mut self, value: Mode)

Sets mode to the provided enum value.

pub fn gid(&self) -> Gid

Returns the enum value of gid, or the default if the field is set to an invalid enum value.

pub fn set_gid(&mut self, value: Gid)

Sets gid to the provided enum value.

pub fn uid(&self) -> Uid

Returns the enum value of uid, or the default if the field is set to an invalid enum value.

pub fn set_uid(&mut self, value: Uid)

Sets uid to the provided enum value.

pub fn acl(&self) -> Acl

Returns the enum value of acl, or the default if the field is set to an invalid enum value.

pub fn set_acl(&mut self, value: Acl)

Sets acl to the provided enum value.

pub fn storage_class(&self) -> StorageClass

Returns the enum value of storage_class, or the default if the field is set to an invalid enum value.

pub fn set_storage_class(&mut self, value: StorageClass)

Sets storage_class to the provided enum value.

pub fn temporary_hold(&self) -> TemporaryHold

Returns the enum value of temporary_hold, or the default if the field is set to an invalid enum value.

pub fn set_temporary_hold(&mut self, value: TemporaryHold)

Sets temporary_hold to the provided enum value.

pub fn kms_key(&self) -> KmsKey

Returns the enum value of kms_key, or the default if the field is set to an invalid enum value.

pub fn set_kms_key(&mut self, value: KmsKey)

Sets kms_key to the provided enum value.

pub fn time_created(&self) -> TimeCreated

Returns the enum value of time_created, or the default if the field is set to an invalid enum value.

pub fn set_time_created(&mut self, value: TimeCreated)

Sets time_created to the provided enum value.