Module google_api_proto::google::streetview::publish::v1

Modules

create_photo_sequence_request

Nested message and enum types in CreatePhotoSequenceRequest.

imu

Nested message and enum types in Imu.

photo

Nested message and enum types in Photo.

photo_sequence

Nested message and enum types in PhotoSequence.

processing_failure_details

Nested message and enum types in ProcessingFailureDetails.

street_view_publish_service_client

Generated client implementations.

upload_ref

Nested message and enum types in UploadRef.

Structs

BatchDeletePhotosRequest

Request to delete multiple [Photos][google.streetview.publish.v1.Photo].

BatchDeletePhotosResponse

Response to batch delete of one or more [Photos][google.streetview.publish.v1.Photo].

BatchGetPhotosRequest

Request to get one or more [Photos][google.streetview.publish.v1.Photo]. By default

BatchGetPhotosResponse

Response to batch get of [Photos][google.streetview.publish.v1.Photo].

BatchUpdatePhotosRequest

Request to update the metadata of photos. Updating the pixels of photos is not supported.

BatchUpdatePhotosResponse

Response to batch update of metadata of one or more [Photos][google.streetview.publish.v1.Photo].

Connection

A connection is the link from a source photo to a destination photo.

CreatePhotoRequest

Request to create a [Photo][google.streetview.publish.v1.Photo].

CreatePhotoSequenceRequest

Request to create a [PhotoSequence][google.streetview.publish.v1.PhotoSequence] from a video.

DeletePhotoRequest

Request to delete a [Photo][google.streetview.publish.v1.Photo].

DeletePhotoSequenceRequest

Request to delete a [PhotoSequence][google.streetview.publish.v1.PhotoSequence].

GetPhotoRequest

Request to get a [Photo][google.streetview.publish.v1.Photo].

GetPhotoSequenceRequest

Request to get a [PhotoSequence][google.streetview.publish.v1.PhotoSequence].

GpsDataGapFailureDetails

Details related to ProcessingFailureReason#GPS_DATA_GAP. If there are multiple GPS data gaps, only the one with the largest duration is reported here.

Imu

IMU data from the device sensors.

ImuDataGapFailureDetails

Details related to ProcessingFailureReason#IMU_DATA_GAP. If there are multiple IMU data gaps, only the one with the largest duration is reported here.

InsufficientGpsFailureDetails

Details related to ProcessingFailureReason#INSUFFICIENT_GPS.

LatLngBounds

A rectangle in geographical coordinates.

Level

Level information containing level number and its corresponding name.

ListPhotoSequencesRequest

Request to list all photo sequences that belong to the user sending the request.

ListPhotoSequencesResponse

Response to list all photo sequences that belong to a user.

ListPhotosRequest

Request to list all photos that belong to the user sending the request.

ListPhotosResponse

Response to list all photos that belong to a user.

NotOutdoorsFailureDetails

Details related to ProcessingFailureReason#NOT_OUTDOORS. If there are multiple indoor frames found, the first frame is recorded here.

Photo

Photo is used to store 360 photos along with photo metadata.

PhotoId

Identifier for a [Photo][google.streetview.publish.v1.Photo].

PhotoResponse

Response payload for a single [Photo][google.streetview.publish.v1.Photo] in batch operations including [BatchGetPhotos][google.streetview.publish.v1.StreetViewPublishService.BatchGetPhotos] and [BatchUpdatePhotos][google.streetview.publish.v1.StreetViewPublishService.BatchUpdatePhotos].

PhotoSequence

A sequence of 360 photos along with metadata.

Place

Place metadata for an entity.

Pose

Raw pose measurement for an entity.

ProcessingFailureDetails

Additional details to accompany the ProcessingFailureReason enum. This message is always expected to be used in conjunction with ProcessingFailureReason, and the oneof value set in this message should match the FailureReason.

UpdatePhotoRequest

Request to update the metadata of a [Photo][google.streetview.publish.v1.Photo]. Updating the pixels of a photo is not supported.

UploadRef

Upload reference for media files.

Enums

PhotoView

Specifies which view of the [Photo][google.streetview.publish.v1.Photo] to include in the response.

ProcessingFailureReason

The possible reasons this [PhotoSequence] [google.streetview.publish.v1.PhotoSequence] failed to process.

ProcessingState

The processing state of the sequence. The states move as follows: