Module google_api_proto::maps::fleetengine::delivery::v1

Modules

delivery_service_client

Generated client implementations.

task

Nested message and enum types in Task.

vehicle_stop

Nested message and enum types in VehicleStop.

Structs

CreateDeliveryVehicleRequest

CreateDeliveryVehicle request message.

CreateTaskRequest

CreateTask request message.

DeliveryVehicle

DeliveryVehicle metadata.

GetDeliveryVehicleRequest

GetDeliveryVehicle request message. Next id: 4

GetTaskRequest

GetTask request message.

ListDeliveryVehiclesRequest

ListDeliveryVehicles request message.

ListDeliveryVehiclesResponse

ListDeliveryVehicles response message.

ListTasksRequest

ListTasks request message.

ListTasksResponse

ListTasks response containing list of Tasks which met the filter criteria in the ListTasksRequest.

LocationInfo

A location with any additional identifiers.

SearchTasksRequest

SearchTasks request message containing the tracking_id used to search Tasks.

SearchTasksResponse

SearchTasks response containing list of Tasks which met the search criteria in the SearchTasksRequest.

Task

A task in the Delivery API represents a single action to track. In general there is a distinction between shipment-related tasks and break tasks. A shipment can have multiple tasks associated with it; for example, one task for the pickup and one for the dropoff or transfer. Different tasks for a given shipment can be handled by different vehicles; for example, one vehicle handles the pickup and drives the shipment to the hub, while another drives the same shipment from the hub to the dropoff.

UpdateDeliveryVehicleRequest

UpdateDeliveryVehicle request message.

UpdateTaskRequest

UpdateTask request message.

VehicleJourneySegment

Represents a vehicle’s travel segment from its previous stop to the current stop. If it is the first active stop, then it is from Vehicle’s current location to this stop.

VehicleStop

Describes a point where a vehicle will stop on its journey to perform one or more tasks.