Struct google_api_proto::maps::fleetengine::v1::SearchVehiclesRequest

pub struct SearchVehiclesRequest {

    pub header: Option<RequestHeader>,
    pub parent: String,
    pub pickup_point: Option<TerminalLocation>,
    pub dropoff_point: Option<TerminalLocation>,
    pub pickup_radius_meters: i32,
    pub count: i32,
    pub minimum_capacity: i32,
    pub trip_types: Vec<i32>,
    pub maximum_staleness: Option<Duration>,
    pub vehicle_types: Vec<VehicleType>,
    pub required_attributes: Vec<VehicleAttribute>,
    pub required_one_of_attributes: Vec<VehicleAttributeList>,
    pub required_one_of_attribute_sets: Vec<VehicleAttributeList>,
    pub order_by: i32,
    pub include_back_to_back: bool,
    pub trip_id: String,
    pub current_trips_present: i32,
    pub filter: String,

}

SearchVehicles request message.

Fields
header: Option<RequestHeader>

The standard Fleet Engine request header.

parent: String

Required. Must be in the format providers/{provider}. The provider must be the Project ID (for example, sample-cloud-project) of the Google Cloud Project of which the service account making this call is a member.

pickup_point: Option<TerminalLocation>

Required. The pickup point to search near.

dropoff_point: Option<TerminalLocation>

The customer’s intended dropoff location. The field is required if trip_types contains TripType.SHARED.

pickup_radius_meters: i32

Required. Defines the vehicle search radius around the pickup point. Only vehicles within the search radius will be returned. Value must be between 400 and 10000 meters (inclusive).

count: i32

Required. Specifies the maximum number of vehicles to return. The value must be between 1 and 50 (inclusive).

minimum_capacity: i32

Required. Specifies the number of passengers being considered for a trip. The value must be greater than or equal to one. The driver is not considered in the capacity value.

trip_types: Vec<i32>

Required. Represents the type of proposed trip. Eligible vehicles are those that can support at least one of the specified trip type.

EXCLUSIVE and SHARED may not be included together. SHARED is not supported when current_trips_present is CURRENT_TRIPS_PRESENT_UNSPECIFIED. UNKNOWN_TRIP_TYPE is not allowed.

maximum_staleness: Option<Duration>

Restricts the search to only those vehicles that have sent location updates to Fleet Engine within the specified duration. Stationary vehicles still transmitting their locations are not considered stale. If this field is not set, the server uses five minutes as the default value.

vehicle_types: Vec<VehicleType>

Required. Restricts the search to vehicles with one of the specified types. At least one vehicle type must be specified. VehicleTypes with a category of UNKNOWN are not allowed.

required_attributes: Vec<VehicleAttribute>

Callers can form complex logical operations using any combination of the required_attributes, required_one_of_attributes, and required_one_of_attribute_sets fields.

required_attributes is a list; required_one_of_attributes uses a message which allows a list of lists. In combination, the two fields allow the composition of this expression:

(required_attributes\[0\] AND required_attributes\[1\] AND ...)
AND
(required_one_of_attributes\[0][0\] OR required_one_of_attributes\[0][1\] OR
...)
AND
(required_one_of_attributes\[1][0\] OR required_one_of_attributes\[1][1\] OR
...)

Restricts the search to only those vehicles with the specified attributes. This field is a conjunction/AND operation. A max of 50 required_attributes is allowed. This matches the maximum number of attributes allowed on a vehicle.

required_one_of_attributes: Vec<VehicleAttributeList>

Restricts the search to only those vehicles with at least one of the specified attributes in each VehicleAttributeList. Within each list, a vehicle must match at least one of the attributes. This field is an inclusive disjunction/OR operation in each VehicleAttributeList and a conjunction/AND operation across the collection of VehicleAttributeList.

required_one_of_attribute_sets: Vec<VehicleAttributeList>

required_one_of_attribute_sets provides additional functionality.

Similar to required_one_of_attributes, required_one_of_attribute_sets uses a message which allows a list of lists, allowing expressions such as this one:

(required_attributes\[0\] AND required_attributes\[1\] AND ...)
AND
(
   (required_one_of_attribute_sets\[0][0\] AND
   required_one_of_attribute_sets\[0][1\] AND
   ...)
   OR
   (required_one_of_attribute_sets\[1][0\] AND
   required_one_of_attribute_sets\[1][1\] AND
   ...)
)

Restricts the search to only those vehicles with all the attributes in a VehicleAttributeList. Within each list, a vehicle must match all of the attributes. This field is a conjunction/AND operation in each VehicleAttributeList and inclusive disjunction/OR operation across the collection of VehicleAttributeList.

order_by: i32

Required. Specifies the desired ordering criterion for results.

include_back_to_back: bool

Indicates if a vehicle with a single active trip is eligible for another match. If false, vehicles with assigned trips are excluded from the search results. If true, search results include vehicles with TripStatus of ENROUTE_TO_DROPOFF.

This field is only considered if a single trip_type of EXCLUSIVE is specified.

The default value is false.

trip_id: String

Indicates the trip associated with this SearchVehicleRequest.

current_trips_present: i32

Restricts vehicles from appearing in the search results based on their current trips.

When current_trips_present is NONE or ANY, trip_types can be either EXCLUSIVE or SHARED, but not both.

filter: String

Optional. A filter query to apply when searching vehicles. See http://aip.dev/160 for examples of the filter syntax.

This field is designed to replace the required_attributes, required_one_of_attributes, and required_one_of_attributes_sets fields. If a non-empty value is specified here, the following fields must be empty: required_attributes, required_one_of_attributes, and required_one_of_attributes_sets.

This filter functions as an AND clause with other constraints, such as minimum_capacity or vehicle_types.

Note that the only queries supported are on vehicle attributes (for example, attributes.<key> = <value> or attributes.<key1> = <value1> AND attributes.<key2> = <value2>). The maximum number of restrictions allowed in a filter query is 50.

Also, all attributes are stored as strings, so the only supported comparisons against attributes are string comparisons. In order to compare against number or boolean values, the values must be explicitly quoted to be treated as strings (for example, attributes.<key> = "10" or attributes.<key> = "true").

Implementations

impl SearchVehiclesRequest

pub fn trip_types( &self ) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<TripType>>

Returns an iterator which yields the valid enum values contained in trip_types.

pub fn push_trip_types(&mut self, value: TripType)

Appends the provided enum value to trip_types.

pub fn order_by(&self) -> VehicleMatchOrder

Returns the enum value of order_by, or the default if the field is set to an invalid enum value.

pub fn set_order_by(&mut self, value: VehicleMatchOrder)

Sets order_by to the provided enum value.

pub fn current_trips_present(&self) -> CurrentTripsPresent

Returns the enum value of current_trips_present, or the default if the field is set to an invalid enum value.

pub fn set_current_trips_present(&mut self, value: CurrentTripsPresent)

Sets current_trips_present to the provided enum value.